Graphics.Vulkan.C.Core10.CommandPool

Documentation

type VkCommandPool = Ptr VkCommandPool_T

VkCommandPool - Opaque handle to a command pool object

See Also

VkCommandBufferAllocateInfo, vkCreateCommandPool, vkDestroyCommandPool, vkFreeCommandBuffers, vkResetCommandPool, vkTrimCommandPool, vkTrimCommandPoolKHR

newtype VkCommandPoolCreateFlagBits

VkCommandPoolCreateFlagBits - Bitmask specifying usage behavior for a command pool

See Also

VkCommandPoolCreateFlags

Constructors

VkCommandPoolCreateFlagBits VkFlags

pattern VK_COMMAND_POOL_CREATE_TRANSIENT_BIT :: VkCommandPoolCreateFlagBits

VK_COMMAND_POOL_CREATE_TRANSIENT_BIT specifies that command buffers allocated from the pool will be short-lived, meaning that they will be reset or freed in a relatively short timeframe. This flag may be used by the implementation to control memory allocation behavior within the pool.

pattern VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT :: VkCommandPoolCreateFlagBits

VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT allows any command buffer allocated from a pool to be individually reset to the initial state; either by calling vkResetCommandBuffer, or via the implicit reset when calling vkBeginCommandBuffer. If this flag is not set on a pool, then vkResetCommandBuffer must not be called for any command buffer allocated from that pool.

type VkCommandPoolCreateFlags = VkCommandPoolCreateFlagBits

VkCommandPoolCreateFlags - Bitmask of VkCommandPoolCreateFlagBits

Description

VkCommandPoolCreateFlags is a bitmask type for setting a mask of zero or more VkCommandPoolCreateFlagBits.

See Also

VkCommandPoolCreateFlagBits, VkCommandPoolCreateInfo

data VkCommandPoolCreateInfo

VkCommandPoolCreateInfo - Structure specifying parameters of a newly created command pool

Valid Usage (Implicit)

See Also

VkCommandPoolCreateFlags, VkStructureType, vkCreateCommandPool

Constructors

VkCommandPoolCreateInfo

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO

vkPNext :: Ptr ()

pNext must be NULL

vkFlags :: VkCommandPoolCreateFlags

flags must be a valid combination of VkCommandPoolCreateFlagBits values

vkQueueFamilyIndex :: Word32

queueFamilyIndex designates a queue family as described in section Queue Family Properties. All command buffers allocated from this command pool must be submitted on queues from the same queue family.

newtype VkCommandPoolResetFlagBits

VkCommandPoolResetFlagBits - Bitmask controlling behavior of a command pool reset

See Also

VkCommandPoolResetFlags

Constructors

VkCommandPoolResetFlagBits VkFlags

pattern VK_COMMAND_POOL_RESET_RELEASE_RESOURCES_BIT :: VkCommandPoolResetFlagBits

VK_COMMAND_POOL_RESET_RELEASE_RESOURCES_BIT specifies that resetting a command pool recycles all of the resources from the command pool back to the system.

type VkCommandPoolResetFlags = VkCommandPoolResetFlagBits

VkCommandPoolResetFlags - Bitmask of VkCommandPoolResetFlagBits

Description

VkCommandPoolResetFlags is a bitmask type for setting a mask of zero or more VkCommandPoolResetFlagBits.

See Also

VkCommandPoolResetFlagBits, vkResetCommandPool

type FN_vkCreateCommandPool = ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkCommandPoolCreateInfo) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pCommandPool" ::: Ptr VkCommandPool) -> IO VkResult

type PFN_vkCreateCommandPool = FunPtr FN_vkCreateCommandPool

vkCreateCommandPool :: DeviceCmds -> ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkCommandPoolCreateInfo) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pCommandPool" ::: Ptr VkCommandPool) -> IO VkResult

vkCreateCommandPool - Create a new command pool object

Parameters

device is the logical device that creates the command pool.

pCreateInfo is a pointer to an instance of the VkCommandPoolCreateInfo structure specifying the state of the command pool object.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pCommandPool points to a VkCommandPool handle in which the created pool is returned.

Valid Usage

pCreateInfo::queueFamilyIndex must be the index of a queue family available in the logical device device.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkCommandPoolCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pCommandPool must be a valid pointer to a VkCommandPool handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkCommandPool, VkCommandPoolCreateInfo, VkDevice

type FN_vkDestroyCommandPool = ("device" ::: VkDevice) -> ("commandPool" ::: VkCommandPool) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> IO ()

type PFN_vkDestroyCommandPool = FunPtr FN_vkDestroyCommandPool

vkDestroyCommandPool :: DeviceCmds -> ("device" ::: VkDevice) -> ("commandPool" ::: VkCommandPool) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> IO ()

vkDestroyCommandPool - Destroy a command pool object

Parameters

device is the logical device that destroys the command pool.

commandPool is the handle of the command pool to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Description

When a pool is destroyed, all command buffers allocated from the pool are freed.

Any primary command buffer allocated from another VkCommandPool that is in the recording or executable state and has a secondary command buffer allocated from commandPool recorded into it, becomes invalid.

Valid Usage

All VkCommandBuffer objects allocated from commandPool must not be in the pending state.

If VkAllocationCallbacks were provided when commandPool was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when commandPool was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid VkDevice handle

If commandPool is not VK_NULL_HANDLE, commandPool must be a valid VkCommandPool handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

If commandPool is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to commandPool must be externally synchronized

See Also

VkAllocationCallbacks, VkCommandPool, VkDevice

type FN_vkResetCommandPool = ("device" ::: VkDevice) -> ("commandPool" ::: VkCommandPool) -> ("flags" ::: VkCommandPoolResetFlags) -> IO VkResult

type PFN_vkResetCommandPool = FunPtr FN_vkResetCommandPool

vkResetCommandPool :: DeviceCmds -> ("device" ::: VkDevice) -> ("commandPool" ::: VkCommandPool) -> ("flags" ::: VkCommandPoolResetFlags) -> IO VkResult

vkResetCommandPool - Reset a command pool

Parameters

device is the logical device that owns the command pool.

commandPool is the command pool to reset.

flags is a bitmask of VkCommandPoolResetFlagBits controlling the reset operation.

Description

Resetting a command pool recycles all of the resources from all of the command buffers allocated from the command pool back to the command pool. All command buffers that have been allocated from the command pool are put in the initial state.

Any primary command buffer allocated from another VkCommandPool that is in the recording or executable state and has a secondary command buffer allocated from commandPool recorded into it, becomes invalid.

Valid Usage

All VkCommandBuffer objects allocated from commandPool must not be in the pending state

Valid Usage (Implicit)

device must be a valid VkDevice handle

commandPool must be a valid VkCommandPool handle

flags must be a valid combination of VkCommandPoolResetFlagBits values

commandPool must have been created, allocated, or retrieved from device

Host Synchronization

Host access to commandPool must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkCommandPool, VkCommandPoolResetFlags, VkDevice