Graphics.Vulkan.C.Core10.ExtensionDiscovery

Documentation

type VK_MAX_EXTENSION_NAME_SIZE = 256

pattern VK_MAX_EXTENSION_NAME_SIZE :: Integral a => a

data VkExtensionProperties

VkExtensionProperties - Structure specifying an extension properties

See Also

vkEnumerateDeviceExtensionProperties, vkEnumerateInstanceExtensionProperties

Constructors

VkExtensionProperties

Fields

vkExtensionName :: Vector VK_MAX_EXTENSION_NAME_SIZE CChar

extensionName is a null-terminated string specifying the name of the extension.

vkSpecVersion :: Word32

specVersion is the version of this extension. It is an integer, incremented with backward compatible changes.

type FN_vkEnumerateDeviceExtensionProperties = ("physicalDevice" ::: VkPhysicalDevice) -> ("pLayerName" ::: Ptr CChar) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkExtensionProperties) -> IO VkResult

type PFN_vkEnumerateDeviceExtensionProperties = FunPtr FN_vkEnumerateDeviceExtensionProperties

vkEnumerateDeviceExtensionProperties :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pLayerName" ::: Ptr CChar) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkExtensionProperties) -> IO VkResult

vkEnumerateDeviceExtensionProperties - Returns properties of available physical device extensions

Parameters

physicalDevice is the physical device that will be queried.

pLayerName is either NULL or a pointer to a null-terminated UTF-8 string naming the layer to retrieve extensions from.

pPropertyCount is a pointer to an integer related to the number of extension properties available or queried, and is treated in the same fashion as the vkEnumerateInstanceExtensionProperties::pPropertyCount parameter.

pProperties is either NULL or a pointer to an array of VkExtensionProperties structures.

Description

When pLayerName parameter is NULL, only extensions provided by the Vulkan implementation or by implicitly enabled layers are returned. When pLayerName is the name of a layer, the device extensions provided by that layer are returned.

Implementations must not advertise any pair of extensions that cannot be enabled together due to behavioral differences, or any extension that cannot be enabled against the advertised version.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

If pLayerName is not NULL, pLayerName must be a null-terminated UTF-8 string

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkExtensionProperties structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_LAYER_NOT_PRESENT

See Also

VkExtensionProperties, VkPhysicalDevice

type FN_vkEnumerateInstanceExtensionProperties = ("pLayerName" ::: Ptr CChar) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkExtensionProperties) -> IO VkResult

type PFN_vkEnumerateInstanceExtensionProperties = FunPtr FN_vkEnumerateInstanceExtensionProperties

vkEnumerateInstanceExtensionProperties :: ("pLayerName" ::: Ptr CChar) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkExtensionProperties) -> IO VkResult