Graphics.Vulkan.C.Core10.Image

Documentation

data VkImageCreateInfo

VkImageCreateInfo - Structure specifying the parameters of a newly created image object

Description

Images created with tiling equal to VK_IMAGE_TILING_LINEAR have further restrictions on their limits and capabilities compared to images created with tiling equal to VK_IMAGE_TILING_OPTIMAL. Creation of images with tiling VK_IMAGE_TILING_LINEAR may not be supported unless other parameters meet all of the constraints:

imageType is VK_IMAGE_TYPE_2D

format is not a depth/stencil format

mipLevels is 1

arrayLayers is 1

samples is VK_SAMPLE_COUNT_1_BIT

usage only includes VK_IMAGE_USAGE_TRANSFER_SRC_BIT and/or VK_IMAGE_USAGE_TRANSFER_DST_BIT

Implementations may support additional limits and capabilities beyond those listed above.

To determine the set of valid usage bits for a given format, call vkGetPhysicalDeviceFormatProperties.

If the size of the resultant image would exceed maxResourceSize, then vkCreateImage must fail and return VK_ERROR_OUT_OF_DEVICE_MEMORY. This failure may occur even when all image creation parameters satisfy their valid usage requirements.

Valid values for some image creation parameters are limited by a numerical upper bound or by inclusion in a bitset. For example, VkImageCreateInfo::arrayLayers is limited by imageCreateMaxArrayLayers, defined below; and VkImageCreateInfo::samples is limited by imageCreateSampleCounts, also defined below.

Several limiting values are defined below, as well as assisting values from which the limiting values are derived. The limiting values are referenced by the relevant valid usage statements of VkImageCreateInfo.

Let VkBool32 imageCreateMaybeLinear indicate if the resultant image may be linear. (The definition below is trivial because certain extensions are disabled in this build of the specification).

If tiling is VK_IMAGE_TILING_LINEAR, then imageCreateMaybeLinear is true.

If tiling is VK_IMAGE_TILING_OPTIMAL, then imageCreateMaybeLinear is false.

Let VkFormatFeatureFlags imageCreateFormatFeatures be the set of format features available during image creation.

If tiling is VK_IMAGE_TILING_LINEAR, then imageCreateFormatFeatures is the value of VkImageFormatProperties::linearTilingFeatures found by calling vkGetPhysicalDeviceFormatProperties with parameter format equal to VkImageCreateInfo::format.

If tiling is VK_IMAGE_TILING_OPTIMAL, then imageCreateFormatFeatures is value of VkImageFormatProperties::optimalTilingFeatures found by calling vkGetPhysicalDeviceFormatProperties with parameter format equal to VkImageCreateInfo::format.

Let uint32_t imageCreateMaxMipLevels be the value of VkImageFormatProperties::maxMipLevels found by calling vkGetPhysicalDeviceImageFormatProperties with parameters format, imageType, tiling, usage, and flags equal to those in VkImageCreateInfo. If vkGetPhysicalDeviceFormatProperties returns an error, then imageCreateMaxMipLevels is undefined.

Let uint32_t imageCreateMaxArrayLayers be defined analogously to imageCreateMaxMipLevels.

Let VkExtent3D imageCreateMaxExtent be defined analogously to imageCreateMaxMipLevels.

Let VkSampleCountFlags imageCreateSampleCounts be defined analogously to imageCreateMaxMipLevels.

Valid Usage

Each of the following values (as described in Image Creation Limits) must not be undefined imageCreateMaxMipLevels, imageCreateMaxArrayLayers, imageCreateMaxExtent, and imageCreateSampleCounts.

If sharingMode is VK_SHARING_MODE_CONCURRENT, pQueueFamilyIndices must be a valid pointer to an array of queueFamilyIndexCount uint32_t values

If sharingMode is VK_SHARING_MODE_CONCURRENT, queueFamilyIndexCount must be greater than 1

If sharingMode is VK_SHARING_MODE_CONCURRENT, each element of pQueueFamilyIndices must be unique and must be less than pQueueFamilyPropertyCount returned by vkGetPhysicalDeviceQueueFamilyProperties for the physicalDevice that was used to create device

format must not be VK_FORMAT_UNDEFINED

extent::width must be greater than 0.

extent::height must be greater than 0.

extent::depth must be greater than 0.

mipLevels must be greater than 0

arrayLayers must be greater than 0

If flags contains VK_IMAGE_CREATE_CUBE_COMPATIBLE_BIT, imageType must be VK_IMAGE_TYPE_2D

extent.width must be less than or equal to imageCreateMaxExtent.width (as defined in Image Creation Limits).

extent.height must be less than or equal to imageCreateMaxExtent.height (as defined in Image Creation Limits).

extent.depth must be less than or equal to imageCreateMaxExtent.depth (as defined in Image Creation Limits).

If imageType is VK_IMAGE_TYPE_2D and flags contains VK_IMAGE_CREATE_CUBE_COMPATIBLE_BIT, extent.width and extent.height must be equal and arrayLayers must be greater than or equal to 6

If imageType is VK_IMAGE_TYPE_1D, both extent.height and extent.depth must be 1

If imageType is VK_IMAGE_TYPE_2D, extent.depth must be 1

mipLevels must be less than or equal to the number of levels in the complete mipmap chain based on extent.width, extent.height, and extent.depth.

mipLevels must be less than or equal to imageCreateMaxMipLevels (as defined in Image Creation Limits).

arrayLayers must be less than or equal to imageCreateMaxArrayLayers (as defined in Image Creation Limits).

If imageType is VK_IMAGE_TYPE_3D, arrayLayers must be 1.

If samples is not VK_SAMPLE_COUNT_1_BIT, then imageType must be VK_IMAGE_TYPE_2D, flags must not contain VK_IMAGE_CREATE_CUBE_COMPATIBLE_BIT, mipLevels must be equal to 1, and imageCreateMaybeLinear (as defined in Image Creation Limits) must be false,

If usage includes VK_IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT, then bits other than VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT, VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT, and VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT must not be set

If usage includes VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT, VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT, VK_IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT, or VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT, extent.width must be less than or equal to VkPhysicalDeviceLimits::maxFramebufferWidth

If usage includes VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT, VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT, VK_IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT, or VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT, extent.height must be less than or equal to VkPhysicalDeviceLimits::maxFramebufferHeight

If usage includes VK_IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT, usage must also contain at least one of VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT, VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT, or VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT.

samples must be a bit value that is set in imageCreateSampleCounts (as defined in Image Creation Limits).

If the multisampled storage images feature is not enabled, and usage contains VK_IMAGE_USAGE_STORAGE_BIT, samples must be VK_SAMPLE_COUNT_1_BIT

If the sparse bindings feature is not enabled, flags must not contain VK_IMAGE_CREATE_SPARSE_BINDING_BIT

If the sparse aliased residency feature is not enabled, flags must not contain VK_IMAGE_CREATE_SPARSE_ALIASED_BIT

If imageType is VK_IMAGE_TYPE_1D, flags must not contain VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT

If the sparse residency for 2D images feature is not enabled, and imageType is VK_IMAGE_TYPE_2D, flags must not contain VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT

If the sparse residency for 3D images feature is not enabled, and imageType is VK_IMAGE_TYPE_3D, flags must not contain VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT

If the sparse residency for images with 2 samples feature is not enabled, imageType is VK_IMAGE_TYPE_2D, and samples is VK_SAMPLE_COUNT_2_BIT, flags must not contain VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT

If the sparse residency for images with 4 samples feature is not enabled, imageType is VK_IMAGE_TYPE_2D, and samples is VK_SAMPLE_COUNT_4_BIT, flags must not contain VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT

If the sparse residency for images with 8 samples feature is not enabled, imageType is VK_IMAGE_TYPE_2D, and samples is VK_SAMPLE_COUNT_8_BIT, flags must not contain VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT

If the sparse residency for images with 16 samples feature is not enabled, imageType is VK_IMAGE_TYPE_2D, and samples is VK_SAMPLE_COUNT_16_BIT, flags must not contain VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT

If flags contains VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT or VK_IMAGE_CREATE_SPARSE_ALIASED_BIT, it must also contain VK_IMAGE_CREATE_SPARSE_BINDING_BIT

If any of the bits VK_IMAGE_CREATE_SPARSE_BINDING_BIT, VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT, or VK_IMAGE_CREATE_SPARSE_ALIASED_BIT are set, VK_IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT must not also be set

initialLayout must be VK_IMAGE_LAYOUT_UNDEFINED or VK_IMAGE_LAYOUT_PREINITIALIZED.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkDedicatedAllocationImageCreateInfoNV, VkExternalFormatANDROID, VkExternalMemoryImageCreateInfo, VkExternalMemoryImageCreateInfoNV, VkImageDrmFormatModifierExplicitCreateInfoEXT, VkImageDrmFormatModifierListCreateInfoEXT, VkImageFormatListCreateInfoKHR, VkImageStencilUsageCreateInfoEXT, or VkImageSwapchainCreateInfoKHR

Each sType member in the pNext chain must be unique

flags must be a valid combination of VkImageCreateFlagBits values

imageType must be a valid VkImageType value

format must be a valid VkFormat value

samples must be a valid VkSampleCountFlagBits value

tiling must be a valid VkImageTiling value

usage must be a valid combination of VkImageUsageFlagBits values

usage must not be 0

sharingMode must be a valid VkSharingMode value

initialLayout must be a valid VkImageLayout value

</section> = See Also

VkExtent3D, VkFormat, VkImageCreateFlags, VkImageLayout, VkImageTiling, VkImageType, VkImageUsageFlags, VkSampleCountFlagBits, VkSharingMode, VkStructureType, vkCreateImage

Constructors

VkImageCreateInfo

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkFlags :: VkImageCreateFlags

flags is a bitmask of VkImageCreateFlagBits describing additional parameters of the image.

vkImageType :: VkImageType

imageType is a VkImageType value specifying the basic dimensionality of the image. Layers in array textures do not count as a dimension for the purposes of the image type.

vkFormat :: VkFormat

format is a VkFormat describing the format and type of the texel blocks that will be contained in the image.

vkExtent :: VkExtent3D

extent is a VkExtent3D describing the number of data elements in each dimension of the base level.

vkMipLevels :: Word32

mipLevels describes the number of levels of detail available for minified sampling of the image.

vkArrayLayers :: Word32

arrayLayers is the number of layers in the image.

vkSamples :: VkSampleCountFlagBits

samples is a VkSampleCountFlagBits specifying the number of samples per texel.

vkTiling :: VkImageTiling

tiling is a VkImageTiling value specifying the tiling arrangement of the texel blocks in memory.

vkUsage :: VkImageUsageFlags

usage is a bitmask of VkImageUsageFlagBits describing the intended usage of the image.

vkSharingMode :: VkSharingMode

sharingMode is a VkSharingMode value specifying the sharing mode of the image when it will be accessed by multiple queue families.

vkQueueFamilyIndexCount :: Word32

queueFamilyIndexCount is the number of entries in the pQueueFamilyIndices array.

vkPQueueFamilyIndices :: Ptr Word32

pQueueFamilyIndices is a list of queue families that will access this image (ignored if sharingMode is not VK_SHARING_MODE_CONCURRENT).

vkInitialLayout :: VkImageLayout

initialLayout is a VkImageLayout value specifying the initial VkImageLayout of all image subresources of the image. See Image Layouts.

newtype VkImageLayout

VkImageLayout - Layout of image and image subresources

Description

The type(s) of device access supported by each layout are:

The layout of each image subresource is not a state of the image subresource itself, but is rather a property of how the data in memory is organized, and thus for each mechanism of accessing an image in the API the application must specify a parameter or structure member that indicates which image layout the image subresource(s) are considered to be in when the image will be accessed. For transfer commands, this is a parameter to the command (see https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#clears and https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#copies). For use as a framebuffer attachment, this is a member in the substructures of the VkRenderPassCreateInfo (see Render Pass). For use in a descriptor set, this is a member in the VkDescriptorImageInfo structure (see https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#descriptorsets-updates).

See Also

VkAttachmentDescription, VkAttachmentDescription2KHR, VkAttachmentReference, VkAttachmentReference2KHR, VkDescriptorImageInfo, VkImageCreateInfo, VkImageMemoryBarrier, vkCmdBindShadingRateImageNV, vkCmdBlitImage, vkCmdClearColorImage, vkCmdClearDepthStencilImage, vkCmdCopyBufferToImage, vkCmdCopyImage, vkCmdCopyImageToBuffer, vkCmdResolveImage

Constructors

VkImageLayout Int32

pattern VK_IMAGE_LAYOUT_UNDEFINED :: VkImageLayout

VK_IMAGE_LAYOUT_UNDEFINED does not support device access. This layout must only be used as the initialLayout member of VkImageCreateInfo or VkAttachmentDescription, or as the oldLayout in an image transition. When transitioning out of this layout, the contents of the memory are not guaranteed to be preserved.

pattern VK_IMAGE_LAYOUT_GENERAL :: VkImageLayout

VK_IMAGE_LAYOUT_GENERAL supports all types of device access.

pattern VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL :: VkImageLayout

VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL must only be used as a color or resolve attachment in a VkFramebuffer. This layout is valid only for image subresources of images created with the VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT usage bit enabled.

pattern VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL :: VkImageLayout

VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL must only be used as a depth/stencil attachment in a VkFramebuffer. This layout is valid only for image subresources of images created with the VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT usage bit enabled.

pattern VK_IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL :: VkImageLayout

VK_IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL must only be used as a read-only depth/stencil attachment in a VkFramebuffer and/or as a read-only image in a shader (which can be read as a sampled image, combined image/sampler and/or input attachment). This layout is valid only for image subresources of images created with the VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT usage bit enabled. Only image views created with a usage value including VK_IMAGE_USAGE_SAMPLED_BIT can be used as a sampled image or combined image/sampler in a shader. Similarly, only image views created with a usage value including VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT can be used as input attachments.

pattern VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL :: VkImageLayout

VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL must only be used as a read-only image in a shader (which can be read as a sampled image, combined image/sampler and/or input attachment). This layout is valid only for image subresources of images created with the VK_IMAGE_USAGE_SAMPLED_BIT or VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT usage bit enabled.

pattern VK_IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL :: VkImageLayout

VK_IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL must only be used as a source image of a transfer command (see the definition of https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#synchronization-pipeline-stages-transfer). This layout is valid only for image subresources of images created with the VK_IMAGE_USAGE_TRANSFER_SRC_BIT usage bit enabled.

pattern VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL :: VkImageLayout

VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL must only be used as a destination image of a transfer command. This layout is valid only for image subresources of images created with the VK_IMAGE_USAGE_TRANSFER_DST_BIT usage bit enabled.

pattern VK_IMAGE_LAYOUT_PREINITIALIZED :: VkImageLayout

VK_IMAGE_LAYOUT_PREINITIALIZED does not support device access. This layout must only be used as the initialLayout member of VkImageCreateInfo or VkAttachmentDescription, or as the oldLayout in an image transition. When transitioning out of this layout, the contents of the memory are preserved. This layout is intended to be used as the initial layout for an image whose contents are written by the host, and hence the data can be written to memory immediately, without first executing a layout transition. Currently, VK_IMAGE_LAYOUT_PREINITIALIZED is only useful with linear images because there is not a standard layout defined for VK_IMAGE_TILING_OPTIMAL images.

data VkSubresourceLayout

VkSubresourceLayout - Structure specifying subresource layout

Description

If the image is linear, then rowPitch, arrayPitch and depthPitch describe the layout of the image subresource in linear memory. For uncompressed formats, rowPitch is the number of bytes between texels with the same x coordinate in adjacent rows (y coordinates differ by one). arrayPitch is the number of bytes between texels with the same x and y coordinate in adjacent array layers of the image (array layer values differ by one). depthPitch is the number of bytes between texels with the same x and y coordinate in adjacent slices of a 3D image (z coordinates differ by one). Expressed as an addressing formula, the starting byte of a texel in the image subresource has address:

// (x,y,z,layer) are in texel coordinates
address(x,y,z,layer) = layer*arrayPitch + z*depthPitch + y*rowPitch + x*elementSize + offset

For compressed formats, the rowPitch is the number of bytes between compressed texel blocks in adjacent rows. arrayPitch is the number of bytes between compressed texel blocks in adjacent array layers. depthPitch is the number of bytes between compressed texel blocks in adjacent slices of a 3D image.

// (x,y,z,layer) are in compressed texel block coordinates
address(x,y,z,layer) = layer*arrayPitch + z*depthPitch + y*rowPitch + x*compressedTexelBlockByteSize + offset;

The value of arrayPitch is undefined for images that were not created as arrays. depthPitch is defined only for 3D images.

If the image has a color format , then the aspectMask member of VkImageSubresource must be VK_IMAGE_ASPECT_COLOR_BIT.

If the image has a depth/stencil format , then aspectMask must be either VK_IMAGE_ASPECT_DEPTH_BIT or VK_IMAGE_ASPECT_STENCIL_BIT. On implementations that store depth and stencil aspects separately, querying each of these image subresource layouts will return a different offset and size representing the region of memory used for that aspect. On implementations that store depth and stencil aspects interleaved, the same offset and size are returned and represent the interleaved memory allocation.

See Also

VkDeviceSize, VkImageDrmFormatModifierExplicitCreateInfoEXT, vkGetImageSubresourceLayout

Constructors

VkSubresourceLayout

Fields

vkOffset :: VkDeviceSize

offset is the byte offset from the start of the image where the image subresource begins.

vkSize :: VkDeviceSize

size is the size in bytes of the image subresource. size includes any extra memory that is required based on rowPitch.

vkRowPitch :: VkDeviceSize

rowPitch describes the number of bytes between each row of texels in an image.

vkArrayPitch :: VkDeviceSize

arrayPitch describes the number of bytes between each array layer of an image.

vkDepthPitch :: VkDeviceSize

depthPitch describes the number of bytes between each slice of 3D image.

type FN_vkCreateImage = ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkImageCreateInfo) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pImage" ::: Ptr VkImage) -> IO VkResult

type PFN_vkCreateImage = FunPtr FN_vkCreateImage

vkCreateImage :: DeviceCmds -> ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkImageCreateInfo) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pImage" ::: Ptr VkImage) -> IO VkResult

vkCreateImage - Create a new image object

Parameters

device is the logical device that creates the image.

pCreateInfo is a pointer to an instance of the VkImageCreateInfo structure containing parameters to be used to create the image.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pImage points to a VkImage handle in which the resulting image object is returned.

Valid Usage

If the flags member of pCreateInfo includes VK_IMAGE_CREATE_SPARSE_BINDING_BIT, creating this VkImage must not cause the total required sparse memory for all currently valid sparse resources on the device to exceed VkPhysicalDeviceLimits::sparseAddressSpaceSize

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkImageCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pImage must be a valid pointer to a VkImage handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDevice, VkImage, VkImageCreateInfo

type FN_vkDestroyImage = ("device" ::: VkDevice) -> ("image" ::: VkImage) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> IO ()

type PFN_vkDestroyImage = FunPtr FN_vkDestroyImage

vkDestroyImage :: DeviceCmds -> ("device" ::: VkDevice) -> ("image" ::: VkImage) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> IO ()

vkDestroyImage - Destroy an image object

Parameters

device is the logical device that destroys the image.

image is the image to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to image, either directly or via a VkImageView, must have completed execution

If VkAllocationCallbacks were provided when image was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when image was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid VkDevice handle

If image is not VK_NULL_HANDLE, image must be a valid VkImage handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

If image is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to image must be externally synchronized

See Also

VkAllocationCallbacks, VkDevice, VkImage

type FN_vkGetImageSubresourceLayout = ("device" ::: VkDevice) -> ("image" ::: VkImage) -> ("pSubresource" ::: Ptr VkImageSubresource) -> ("pLayout" ::: Ptr VkSubresourceLayout) -> IO ()

type PFN_vkGetImageSubresourceLayout = FunPtr FN_vkGetImageSubresourceLayout

vkGetImageSubresourceLayout :: DeviceCmds -> ("device" ::: VkDevice) -> ("image" ::: VkImage) -> ("pSubresource" ::: Ptr VkImageSubresource) -> ("pLayout" ::: Ptr VkSubresourceLayout) -> IO ()

vkGetImageSubresourceLayout - Retrieve information about an image subresource

Parameters

device is the logical device that owns the image.

image is the image whose layout is being queried.

pSubresource is a pointer to a VkImageSubresource structure selecting a specific image for the image subresource.

pLayout points to a VkSubresourceLayout structure in which the layout is returned.

Description

The image must be linear. The returned layout is valid for host access.

vkGetImageSubresourceLayout is invariant for the lifetime of a single image.

Valid Usage

image must have been created with tiling equal to VK_IMAGE_TILING_LINEAR

The aspectMask member of pSubresource must only have a single bit set

The mipLevel member of pSubresource must be less than the mipLevels specified in VkImageCreateInfo when image was created

The arrayLayer member of pSubresource must be less than the arrayLayers specified in VkImageCreateInfo when image was created

Valid Usage (Implicit)

device must be a valid VkDevice handle

image must be a valid VkImage handle

pSubresource must be a valid pointer to a valid VkImageSubresource structure

pLayout must be a valid pointer to a VkSubresourceLayout structure

image must have been created, allocated, or retrieved from device

See Also

VkDevice, VkImage, VkImageSubresource, VkSubresourceLayout