Graphics.Vulkan.C.Core10.LayerDiscovery

Documentation

type VK_MAX_DESCRIPTION_SIZE = 256

pattern VK_MAX_DESCRIPTION_SIZE :: Integral a => a

data VkLayerProperties

VkLayerProperties - Structure specifying layer properties

See Also

vkEnumerateDeviceLayerProperties, vkEnumerateInstanceLayerProperties

Constructors

VkLayerProperties

Fields

vkLayerName :: Vector VK_MAX_EXTENSION_NAME_SIZE CChar

layerName is a null-terminated UTF-8 string specifying the name of the layer. Use this name in the ppEnabledLayerNames array passed in the VkInstanceCreateInfo structure to enable this layer for an instance.

vkSpecVersion :: Word32

specVersion is the Vulkan version the layer was written to, encoded as described in https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#extendingvulkan-coreversions-versionnumbers.

vkImplementationVersion :: Word32

implementationVersion is the version of this layer. It is an integer, increasing with backward compatible changes.

vkDescription :: Vector VK_MAX_DESCRIPTION_SIZE CChar

description is a null-terminated UTF-8 string providing additional details that can be used by the application to identify the layer.

type FN_vkEnumerateDeviceLayerProperties = ("physicalDevice" ::: VkPhysicalDevice) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkLayerProperties) -> IO VkResult

type PFN_vkEnumerateDeviceLayerProperties = FunPtr FN_vkEnumerateDeviceLayerProperties

vkEnumerateDeviceLayerProperties :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkLayerProperties) -> IO VkResult

vkEnumerateDeviceLayerProperties - Returns properties of available physical device layers

Parameters

pPropertyCount is a pointer to an integer related to the number of layer properties available or queried.

pProperties is either NULL or a pointer to an array of VkLayerProperties structures.

Description

If pProperties is NULL, then the number of layer properties available is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If pPropertyCount is less than the number of layer properties available, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of layers available, VK_INCOMPLETE will be returned instead of VK_SUCCESS, to indicate that not all the available layer properties were returned.

The list of layers enumerated by vkEnumerateDeviceLayerProperties must be exactly the sequence of layers enabled for the instance. The members of VkLayerProperties for each enumerated layer must be the same as the properties when the layer was enumerated by vkEnumerateInstanceLayerProperties.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkLayerProperties structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkLayerProperties, VkPhysicalDevice

type FN_vkEnumerateInstanceLayerProperties = ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkLayerProperties) -> IO VkResult

type PFN_vkEnumerateInstanceLayerProperties = FunPtr FN_vkEnumerateInstanceLayerProperties

vkEnumerateInstanceLayerProperties :: ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkLayerProperties) -> IO VkResult