Graphics.Vulkan.C.Core10.MemoryManagement

Documentation

type VkBuffer = Ptr VkBuffer_T

VkBuffer - Opaque handle to a buffer object

See Also

VkBindBufferMemoryInfo, VkBufferDeviceAddressInfoEXT, VkBufferMemoryBarrier, VkBufferMemoryRequirementsInfo2, VkBufferViewCreateInfo, VkCmdProcessCommandsInfoNVX, VkConditionalRenderingBeginInfoEXT, VkDedicatedAllocationMemoryAllocateInfoNV, VkDescriptorBufferInfo, VkGeometryAABBNV, VkGeometryTrianglesNV, VkIndirectCommandsTokenNVX, VkMemoryDedicatedAllocateInfo, VkObjectTableIndexBufferEntryNVX, VkObjectTableVertexBufferEntryNVX, VkSparseBufferMemoryBindInfo, vkBindBufferMemory, vkCmdBeginTransformFeedbackEXT, vkCmdBindIndexBuffer, vkCmdBindTransformFeedbackBuffersEXT, vkCmdBindVertexBuffers, vkCmdBuildAccelerationStructureNV, vkCmdCopyBuffer, vkCmdCopyBufferToImage, vkCmdCopyImageToBuffer, vkCmdCopyQueryPoolResults, vkCmdDispatchIndirect, vkCmdDrawIndexedIndirect, vkCmdDrawIndexedIndirectCountAMD, vkCmdDrawIndexedIndirectCountKHR, vkCmdDrawIndirect, vkCmdDrawIndirectByteCountEXT, vkCmdDrawIndirectCountAMD, vkCmdDrawIndirectCountKHR, vkCmdDrawMeshTasksIndirectCountNV, vkCmdDrawMeshTasksIndirectNV, vkCmdEndTransformFeedbackEXT, vkCmdFillBuffer, vkCmdTraceRaysNV, vkCmdUpdateBuffer, vkCmdWriteBufferMarkerAMD, vkCreateBuffer, vkDestroyBuffer, vkGetBufferMemoryRequirements

type VkImage = Ptr VkImage_T

VkImage - Opaque handle to an image object

See Also

VkBindImageMemoryInfo, VkDedicatedAllocationMemoryAllocateInfoNV, VkImageMemoryBarrier, VkImageMemoryRequirementsInfo2, VkImageSparseMemoryRequirementsInfo2, VkImageViewCreateInfo, VkMemoryDedicatedAllocateInfo, VkSparseImageMemoryBindInfo, VkSparseImageOpaqueMemoryBindInfo, vkBindImageMemory, vkCmdBlitImage, vkCmdClearColorImage, vkCmdClearDepthStencilImage, vkCmdCopyBufferToImage, vkCmdCopyImage, vkCmdCopyImageToBuffer, vkCmdResolveImage, vkCreateImage, vkDestroyImage, vkGetImageDrmFormatModifierPropertiesEXT, vkGetImageMemoryRequirements, vkGetImageSparseMemoryRequirements, vkGetImageSubresourceLayout, vkGetSwapchainImagesKHR

data VkMemoryRequirements

VkMemoryRequirements - Structure specifying memory requirements

See Also

VkDeviceSize, VkMemoryRequirements2, vkGetBufferMemoryRequirements, vkGetImageMemoryRequirements

Constructors

VkMemoryRequirements

Fields

vkSize :: VkDeviceSize

size is the size, in bytes, of the memory allocation required for the resource.

vkAlignment :: VkDeviceSize

alignment is the alignment, in bytes, of the offset within the allocation required for the resource.

vkMemoryTypeBits :: Word32

memoryTypeBits is a bitmask and contains one bit set for every supported memory type for the resource. Bit i is set if and only if the memory type i in the VkPhysicalDeviceMemoryProperties structure for the physical device is supported for the resource.

type FN_vkBindBufferMemory = ("device" ::: VkDevice) -> ("buffer" ::: VkBuffer) -> ("memory" ::: VkDeviceMemory) -> ("memoryOffset" ::: VkDeviceSize) -> IO VkResult

type PFN_vkBindBufferMemory = FunPtr FN_vkBindBufferMemory

vkBindBufferMemory :: DeviceCmds -> ("device" ::: VkDevice) -> ("buffer" ::: VkBuffer) -> ("memory" ::: VkDeviceMemory) -> ("memoryOffset" ::: VkDeviceSize) -> IO VkResult

vkBindBufferMemory - Bind device memory to a buffer object

Parameters

device is the logical device that owns the buffer and memory.

buffer is the buffer to be attached to memory.

memory is a VkDeviceMemory object describing the device memory to attach.

memoryOffset is the start offset of the region of memory which is to be bound to the buffer. The number of bytes returned in the VkMemoryRequirements::size member in memory, starting from memoryOffset bytes, will be bound to the specified buffer.

Valid Usage

buffer must not already be backed by a memory object

buffer must not have been created with any sparse memory binding flags

memoryOffset must be less than the size of memory

memory must have been allocated using one of the memory types allowed in the memoryTypeBits member of the VkMemoryRequirements structure returned from a call to vkGetBufferMemoryRequirements with buffer

memoryOffset must be an integer multiple of the alignment member of the VkMemoryRequirements structure returned from a call to vkGetBufferMemoryRequirements with buffer

The size member of the VkMemoryRequirements structure returned from a call to vkGetBufferMemoryRequirements with buffer must be less than or equal to the size of memory minus memoryOffset

Valid Usage (Implicit)

device must be a valid VkDevice handle

buffer must be a valid VkBuffer handle

memory must be a valid VkDeviceMemory handle

buffer must have been created, allocated, or retrieved from device

memory must have been created, allocated, or retrieved from device

Host Synchronization

Host access to buffer must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkBuffer, VkDevice, VkDeviceMemory, VkDeviceSize

type FN_vkBindImageMemory = ("device" ::: VkDevice) -> ("image" ::: VkImage) -> ("memory" ::: VkDeviceMemory) -> ("memoryOffset" ::: VkDeviceSize) -> IO VkResult

type PFN_vkBindImageMemory = FunPtr FN_vkBindImageMemory

vkBindImageMemory :: DeviceCmds -> ("device" ::: VkDevice) -> ("image" ::: VkImage) -> ("memory" ::: VkDeviceMemory) -> ("memoryOffset" ::: VkDeviceSize) -> IO VkResult

vkBindImageMemory - Bind device memory to an image object

Parameters

device is the logical device that owns the image and memory.

image is the image.

memory is the VkDeviceMemory object describing the device memory to attach.

memoryOffset is the start offset of the region of memory which is to be bound to the image. The number of bytes returned in the VkMemoryRequirements::size member in memory, starting from memoryOffset bytes, will be bound to the specified image.

Valid Usage

image must not already be backed by a memory object

image must not have been created with any sparse memory binding flags

memoryOffset must be less than the size of memory

memory must have been allocated using one of the memory types allowed in the memoryTypeBits member of the VkMemoryRequirements structure returned from a call to vkGetImageMemoryRequirements with image

memoryOffset must be an integer multiple of the alignment member of the VkMemoryRequirements structure returned from a call to vkGetImageMemoryRequirements with image

The size member of the VkMemoryRequirements structure returned from a call to vkGetImageMemoryRequirements with image must be less than or equal to the size of memory minus memoryOffset

Valid Usage (Implicit)

device must be a valid VkDevice handle

image must be a valid VkImage handle

memory must be a valid VkDeviceMemory handle

image must have been created, allocated, or retrieved from device

memory must have been created, allocated, or retrieved from device

Host Synchronization

Host access to image must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDevice, VkDeviceMemory, VkDeviceSize, VkImage

type FN_vkGetBufferMemoryRequirements = ("device" ::: VkDevice) -> ("buffer" ::: VkBuffer) -> ("pMemoryRequirements" ::: Ptr VkMemoryRequirements) -> IO ()

type PFN_vkGetBufferMemoryRequirements = FunPtr FN_vkGetBufferMemoryRequirements

vkGetBufferMemoryRequirements :: DeviceCmds -> ("device" ::: VkDevice) -> ("buffer" ::: VkBuffer) -> ("pMemoryRequirements" ::: Ptr VkMemoryRequirements) -> IO ()

vkGetBufferMemoryRequirements - Returns the memory requirements for specified Vulkan object

Parameters

device is the logical device that owns the buffer.

buffer is the buffer to query.

pMemoryRequirements points to an instance of the VkMemoryRequirements structure in which the memory requirements of the buffer object are returned.

Valid Usage (Implicit)

See Also

VkBuffer, VkDevice, VkMemoryRequirements

type FN_vkGetImageMemoryRequirements = ("device" ::: VkDevice) -> ("image" ::: VkImage) -> ("pMemoryRequirements" ::: Ptr VkMemoryRequirements) -> IO ()

type PFN_vkGetImageMemoryRequirements = FunPtr FN_vkGetImageMemoryRequirements

vkGetImageMemoryRequirements :: DeviceCmds -> ("device" ::: VkDevice) -> ("image" ::: VkImage) -> ("pMemoryRequirements" ::: Ptr VkMemoryRequirements) -> IO ()

vkGetImageMemoryRequirements - Returns the memory requirements for specified Vulkan object

Parameters

device is the logical device that owns the image.

image is the image to query.

pMemoryRequirements points to an instance of the VkMemoryRequirements structure in which the memory requirements of the image object are returned.

Valid Usage (Implicit)

See Also

VkDevice, VkImage, VkMemoryRequirements