Graphics.Vulkan.C.Core10.Pass

Documentation

newtype VkAccessFlagBits

VkAccessFlagBits - Bitmask specifying memory access types that will participate in a memory dependency

Description

Certain access types are only performed by a subset of pipeline stages. Any synchronization command that takes both stage masks and access masks uses both to define the access scopes - only the specified access types performed by the specified stages are included in the access scope. An application must not specify an access flag in a synchronization command if it does not include a pipeline stage in the corresponding stage mask that is able to perform accesses of that type. The following table lists, for each access flag, which pipeline stages can perform that type of access.

+-----------------------------------+-----------------------------------+
| Access flag                       | Supported pipeline stages         |
+===================================+===================================+
| 'VK_ACCESS_INDIRECT_COMMAND_READ_ | 'Graphics.Vulkan.C.Core10.Queue.V |
| BIT'                              | K_PIPELINE_STAGE_DRAW_INDIRECT_BI |
|                                   | T'                                |
+-----------------------------------+-----------------------------------+
| 'VK_ACCESS_INDEX_READ_BIT'        | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_VERTEX_INPUT_BIT |
|                                   | '                                 |
+-----------------------------------+-----------------------------------+
| 'VK_ACCESS_VERTEX_ATTRIBUTE_READ_ | 'Graphics.Vulkan.C.Core10.Queue.V |
| BIT'                              | K_PIPELINE_STAGE_VERTEX_INPUT_BIT |
|                                   | '                                 |
+-----------------------------------+-----------------------------------+
| 'VK_ACCESS_UNIFORM_READ_BIT'      | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_VERTEX_SHADER_BI |
|                                   | T',                               |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_TESSELLATION_CON |
|                                   | TROL_SHADER_BIT',                 |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_TESSELLATION_EVA |
|                                   | LUATION_SHADER_BIT',              |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_GEOMETRY_SHADER_ |
|                                   | BIT',                             |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_FRAGMENT_SHADER_ |
|                                   | BIT',                             |
|                                   | or                                |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_COMPUTE_SHADER_B |
|                                   | IT'                               |
+-----------------------------------+-----------------------------------+
| 'VK_ACCESS_SHADER_READ_BIT'       | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_VERTEX_SHADER_BI |
|                                   | T',                               |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_TESSELLATION_CON |
|                                   | TROL_SHADER_BIT',                 |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_TESSELLATION_EVA |
|                                   | LUATION_SHADER_BIT',              |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_GEOMETRY_SHADER_ |
|                                   | BIT',                             |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_FRAGMENT_SHADER_ |
|                                   | BIT',                             |
|                                   | or                                |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_COMPUTE_SHADER_B |
|                                   | IT'                               |
+-----------------------------------+-----------------------------------+
| 'VK_ACCESS_SHADER_WRITE_BIT'      | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_VERTEX_SHADER_BI |
|                                   | T',                               |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_TESSELLATION_CON |
|                                   | TROL_SHADER_BIT',                 |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_TESSELLATION_EVA |
|                                   | LUATION_SHADER_BIT',              |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_GEOMETRY_SHADER_ |
|                                   | BIT',                             |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_FRAGMENT_SHADER_ |
|                                   | BIT',                             |
|                                   | or                                |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_COMPUTE_SHADER_B |
|                                   | IT'                               |
+-----------------------------------+-----------------------------------+
| 'VK_ACCESS_INPUT_ATTACHMENT_READ_ | 'Graphics.Vulkan.C.Core10.Queue.V |
| BIT'                              | K_PIPELINE_STAGE_FRAGMENT_SHADER_ |
|                                   | BIT'                              |
+-----------------------------------+-----------------------------------+
| 'VK_ACCESS_COLOR_ATTACHMENT_READ_ | 'Graphics.Vulkan.C.Core10.Queue.V |
| BIT'                              | K_PIPELINE_STAGE_COLOR_ATTACHMENT |
|                                   | _OUTPUT_BIT'                      |
+-----------------------------------+-----------------------------------+
| 'VK_ACCESS_COLOR_ATTACHMENT_WRITE | 'Graphics.Vulkan.C.Core10.Queue.V |
| _BIT'                             | K_PIPELINE_STAGE_COLOR_ATTACHMENT |
|                                   | _OUTPUT_BIT'                      |
+-----------------------------------+-----------------------------------+
| 'VK_ACCESS_DEPTH_STENCIL_ATTACHME | 'Graphics.Vulkan.C.Core10.Queue.V |
| NT_READ_BIT'                      | K_PIPELINE_STAGE_EARLY_FRAGMENT_T |
|                                   | ESTS_BIT',                        |
|                                   | or                                |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_LATE_FRAGMENT_TE |
|                                   | STS_BIT'                          |
+-----------------------------------+-----------------------------------+
| 'VK_ACCESS_DEPTH_STENCIL_ATTACHME | 'Graphics.Vulkan.C.Core10.Queue.V |
| NT_WRITE_BIT'                     | K_PIPELINE_STAGE_EARLY_FRAGMENT_T |
|                                   | ESTS_BIT',                        |
|                                   | or                                |
|                                   | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_LATE_FRAGMENT_TE |
|                                   | STS_BIT'                          |
+-----------------------------------+-----------------------------------+
| 'VK_ACCESS_TRANSFER_READ_BIT'     | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_TRANSFER_BIT'    |
+-----------------------------------+-----------------------------------+
| 'VK_ACCESS_TRANSFER_WRITE_BIT'    | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_TRANSFER_BIT'    |
+-----------------------------------+-----------------------------------+
| 'VK_ACCESS_HOST_READ_BIT'         | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_HOST_BIT'        |
+-----------------------------------+-----------------------------------+
| 'VK_ACCESS_HOST_WRITE_BIT'        | 'Graphics.Vulkan.C.Core10.Queue.V |
|                                   | K_PIPELINE_STAGE_HOST_BIT'        |
+-----------------------------------+-----------------------------------+
| 'VK_ACCESS_MEMORY_READ_BIT'       | N\/A                              |
+-----------------------------------+-----------------------------------+
| 'VK_ACCESS_MEMORY_WRITE_BIT'      | N\/A                              |
+-----------------------------------+-----------------------------------+

Supported access types

If a memory object does not have the VK_MEMORY_PROPERTY_HOST_COHERENT_BIT property, then vkFlushMappedMemoryRanges must be called in order to guarantee that writes to the memory object from the host are made available to the host domain, where they can be further made available to the device domain via a domain operation. Similarly, vkInvalidateMappedMemoryRanges must be called to guarantee that writes which are available to the host domain are made visible to host operations.

If the memory object does have the VK_MEMORY_PROPERTY_HOST_COHERENT_BIT property flag, writes to the memory object from the host are automatically made available to the host domain. Similarly, writes made available to the host domain are automatically made visible to the host.

Note

The vkQueueSubmit command automatically performs a domain operation from host to device for all writes performed before the command executes, so in most cases an explicit memory barrier is not needed for this case. In the few circumstances where a submit does not occur between the host write and the device read access, writes can be made available by using an explicit memory barrier.

See Also

VkAccessFlags

Constructors

VkAccessFlagBits VkFlags

pattern VK_ACCESS_INDIRECT_COMMAND_READ_BIT :: VkAccessFlagBits

VK_ACCESS_INDIRECT_COMMAND_READ_BIT specifies read access to indirect command data read as part of an indirect drawing or dispatch command.

pattern VK_ACCESS_INDEX_READ_BIT :: VkAccessFlagBits

VK_ACCESS_INDEX_READ_BIT specifies read access to an index buffer as part of an indexed drawing command, bound by vkCmdBindIndexBuffer.

pattern VK_ACCESS_VERTEX_ATTRIBUTE_READ_BIT :: VkAccessFlagBits

VK_ACCESS_VERTEX_ATTRIBUTE_READ_BIT specifies read access to a vertex buffer as part of a drawing command, bound by vkCmdBindVertexBuffers.

pattern VK_ACCESS_UNIFORM_READ_BIT :: VkAccessFlagBits

VK_ACCESS_UNIFORM_READ_BIT specifies read access to a uniform buffer.

pattern VK_ACCESS_INPUT_ATTACHMENT_READ_BIT :: VkAccessFlagBits

VK_ACCESS_INPUT_ATTACHMENT_READ_BIT specifies read access to an input attachment within a render pass during fragment shading.

pattern VK_ACCESS_SHADER_READ_BIT :: VkAccessFlagBits

VK_ACCESS_SHADER_READ_BIT specifies read access to a storage buffer, uniform texel buffer, storage texel buffer, sampled image, or storage image.

pattern VK_ACCESS_SHADER_WRITE_BIT :: VkAccessFlagBits

VK_ACCESS_SHADER_WRITE_BIT specifies write access to a storage buffer, storage texel buffer, or storage image.

pattern VK_ACCESS_COLOR_ATTACHMENT_READ_BIT :: VkAccessFlagBits

VK_ACCESS_COLOR_ATTACHMENT_READ_BIT specifies read access to a color attachment, such as via blending, logic operations, or via certain subpass load operations.

pattern VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT :: VkAccessFlagBits

VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT specifies write access to a color or resolve attachment during a render pass or via certain subpass load and store operations.

pattern VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_READ_BIT :: VkAccessFlagBits

VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_READ_BIT specifies read access to a depth/stencil attachment, via depth or stencil operations or via certain subpass load operations.

pattern VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT :: VkAccessFlagBits

VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT specifies write access to a depth/stencil attachment, via depth or stencil operations or via certain subpass load and store operations.

pattern VK_ACCESS_TRANSFER_READ_BIT :: VkAccessFlagBits

VK_ACCESS_TRANSFER_READ_BIT specifies read access to an image or buffer in a copy operation.

pattern VK_ACCESS_TRANSFER_WRITE_BIT :: VkAccessFlagBits

VK_ACCESS_TRANSFER_WRITE_BIT specifies write access to an image or buffer in a clear or copy operation.

pattern VK_ACCESS_HOST_READ_BIT :: VkAccessFlagBits

VK_ACCESS_HOST_READ_BIT specifies read access by a host operation. Accesses of this type are not performed through a resource, but directly on memory.

pattern VK_ACCESS_HOST_WRITE_BIT :: VkAccessFlagBits

VK_ACCESS_HOST_WRITE_BIT specifies write access by a host operation. Accesses of this type are not performed through a resource, but directly on memory.

pattern VK_ACCESS_MEMORY_READ_BIT :: VkAccessFlagBits

VK_ACCESS_MEMORY_READ_BIT specifies read access via non-specific entities. These entities include the Vulkan device and host, but may also include entities external to the Vulkan device or otherwise not part of the core Vulkan pipeline. When included in a destination access mask, makes all available writes visible to all future read accesses on entities known to the Vulkan device.

pattern VK_ACCESS_MEMORY_WRITE_BIT :: VkAccessFlagBits

VK_ACCESS_MEMORY_WRITE_BIT specifies write access via non-specific entities. These entities include the Vulkan device and host, but may also include entities external to the Vulkan device or otherwise not part of the core Vulkan pipeline. When included in a source access mask, all writes that are performed by entities known to the Vulkan device are made available. When included in a destination access mask, makes all available writes visible to all future write accesses on entities known to the Vulkan device.

type VkAccessFlags = VkAccessFlagBits

VkAccessFlags - Bitmask of VkAccessFlagBits

Description

VkAccessFlags is a bitmask type for setting a mask of zero or more VkAccessFlagBits.

See Also

VkAccessFlagBits, VkBufferMemoryBarrier, VkImageMemoryBarrier, VkMemoryBarrier, VkSubpassDependency, VkSubpassDependency2KHR

data VkAttachmentDescription

VkAttachmentDescription - Structure specifying an attachment description

Description

If the attachment uses a color format, then loadOp and storeOp are used, and stencilLoadOp and stencilStoreOp are ignored. If the format has depth and/or stencil components, loadOp and storeOp apply only to the depth data, while stencilLoadOp and stencilStoreOp define how the stencil data is handled. loadOp and stencilLoadOp define the load operations that execute as part of the first subpass that uses the attachment. storeOp and stencilStoreOp define the store operations that execute as part of the last subpass that uses the attachment.

The load operation for each sample in an attachment happens-before any recorded command which accesses the sample in the first subpass where the attachment is used. Load operations for attachments with a depth/stencil format execute in the VK_PIPELINE_STAGE_EARLY_FRAGMENT_TESTS_BIT pipeline stage. Load operations for attachments with a color format execute in the VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT pipeline stage.

The store operation for each sample in an attachment happens-after any recorded command which accesses the sample in the last subpass where the attachment is used. Store operations for attachments with a depth/stencil format execute in the VK_PIPELINE_STAGE_LATE_FRAGMENT_TESTS_BIT pipeline stage. Store operations for attachments with a color format execute in the VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT pipeline stage.

If an attachment is not used by any subpass, then loadOp, storeOp, stencilStoreOp, and stencilLoadOp are ignored, and the attachment’s memory contents will not be modified by execution of a render pass instance.

During a render pass instance, input/color attachments with color formats that have a component size of 8, 16, or 32 bits must be represented in the attachment’s format throughout the instance. Attachments with other floating- or fixed-point color formats, or with depth components may be represented in a format with a precision higher than the attachment format, but must be represented with the same range. When such a component is loaded via the loadOp, it will be converted into an implementation-dependent format used by the render pass. Such components must be converted from the render pass format, to the format of the attachment, before they are resolved or stored at the end of a render pass instance via storeOp. Conversions occur as described in Numeric Representation and Computation and Fixed-Point Data Conversions.

If flags includes VK_ATTACHMENT_DESCRIPTION_MAY_ALIAS_BIT, then the attachment is treated as if it shares physical memory with another attachment in the same render pass. This information limits the ability of the implementation to reorder certain operations (like layout transitions and the loadOp) such that it is not improperly reordered against other uses of the same physical memory via a different attachment. This is described in more detail below.

Valid Usage (Implicit)

See Also

VkAttachmentDescriptionFlags, VkAttachmentLoadOp, VkAttachmentStoreOp, VkFormat, VkImageLayout, VkRenderPassCreateInfo, VkSampleCountFlagBits

Constructors

VkAttachmentDescription

Fields

vkFlags :: VkAttachmentDescriptionFlags

flags must be a valid combination of VkAttachmentDescriptionFlagBits values

vkFormat :: VkFormat

format must be a valid VkFormat value

vkSamples :: VkSampleCountFlagBits

samples must be a valid VkSampleCountFlagBits value

vkLoadOp :: VkAttachmentLoadOp

loadOp must be a valid VkAttachmentLoadOp value

vkStoreOp :: VkAttachmentStoreOp

storeOp must be a valid VkAttachmentStoreOp value

vkStencilLoadOp :: VkAttachmentLoadOp

stencilLoadOp must be a valid VkAttachmentLoadOp value

vkStencilStoreOp :: VkAttachmentStoreOp

stencilStoreOp must be a valid VkAttachmentStoreOp value

vkInitialLayout :: VkImageLayout

initialLayout must be a valid VkImageLayout value

vkFinalLayout :: VkImageLayout

finalLayout must be a valid VkImageLayout value

newtype VkAttachmentDescriptionFlagBits

VkAttachmentDescriptionFlagBits - Bitmask specifying additional properties of an attachment

See Also

VkAttachmentDescriptionFlags

Constructors

VkAttachmentDescriptionFlagBits VkFlags

pattern VK_ATTACHMENT_DESCRIPTION_MAY_ALIAS_BIT :: VkAttachmentDescriptionFlagBits

VK_ATTACHMENT_DESCRIPTION_MAY_ALIAS_BIT specifies that the attachment aliases the same device memory as other attachments.

type VkAttachmentDescriptionFlags = VkAttachmentDescriptionFlagBits

VkAttachmentDescriptionFlags - Bitmask of VkAttachmentDescriptionFlagBits

Description

VkAttachmentDescriptionFlags is a bitmask type for setting a mask of zero or more VkAttachmentDescriptionFlagBits.

See Also

VkAttachmentDescription, VkAttachmentDescription2KHR, VkAttachmentDescriptionFlagBits

newtype VkAttachmentLoadOp

VkAttachmentLoadOp - Specify how contents of an attachment are treated at the beginning of a subpass

See Also

VkAttachmentDescription, VkAttachmentDescription2KHR

Constructors

VkAttachmentLoadOp Int32

pattern VK_ATTACHMENT_LOAD_OP_LOAD :: VkAttachmentLoadOp

VK_ATTACHMENT_LOAD_OP_LOAD specifies that the previous contents of the image within the render area will be preserved. For attachments with a depth/stencil format, this uses the access type VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_READ_BIT. For attachments with a color format, this uses the access type VK_ACCESS_COLOR_ATTACHMENT_READ_BIT.

pattern VK_ATTACHMENT_LOAD_OP_CLEAR :: VkAttachmentLoadOp

VK_ATTACHMENT_LOAD_OP_CLEAR specifies that the contents within the render area will be cleared to a uniform value, which is specified when a render pass instance is begun. For attachments with a depth/stencil format, this uses the access type VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT. For attachments with a color format, this uses the access type VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT.

pattern VK_ATTACHMENT_LOAD_OP_DONT_CARE :: VkAttachmentLoadOp

VK_ATTACHMENT_LOAD_OP_DONT_CARE specifies that the previous contents within the area need not be preserved; the contents of the attachment will be undefined inside the render area. For attachments with a depth/stencil format, this uses the access type VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT. For attachments with a color format, this uses the access type VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT.

data VkAttachmentReference

VkAttachmentReference - Structure specifying an attachment reference

Valid Usage

If attachment is not VK_ATTACHMENT_UNUSED, layout must not be VK_IMAGE_LAYOUT_UNDEFINED or VK_IMAGE_LAYOUT_PREINITIALIZED

Valid Usage (Implicit)

layout must be a valid VkImageLayout value

See Also

VkImageLayout, VkRenderPassFragmentDensityMapCreateInfoEXT, VkSubpassDescription

Constructors

VkAttachmentReference

Fields

vkAttachment :: Word32

attachment is either an integer value identifying an attachment at the corresponding index in VkRenderPassCreateInfo::pAttachments, or VK_ATTACHMENT_UNUSED to signify that this attachment is not used.

vkLayout :: VkImageLayout

layout is a VkImageLayout value specifying the layout the attachment uses during the subpass.

newtype VkAttachmentStoreOp

VkAttachmentStoreOp - Specify how contents of an attachment are treated at the end of a subpass

See Also

VkAttachmentDescription, VkAttachmentDescription2KHR

Constructors

VkAttachmentStoreOp Int32

pattern VK_ATTACHMENT_STORE_OP_STORE :: VkAttachmentStoreOp

VK_ATTACHMENT_STORE_OP_STORE specifies the contents generated during the render pass and within the render area are written to memory. For attachments with a depth/stencil format, this uses the access type VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT. For attachments with a color format, this uses the access type VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT.

pattern VK_ATTACHMENT_STORE_OP_DONT_CARE :: VkAttachmentStoreOp

VK_ATTACHMENT_STORE_OP_DONT_CARE specifies the contents within the render area are not needed after rendering, and may be discarded; the contents of the attachment will be undefined inside the render area. For attachments with a depth/stencil format, this uses the access type VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT. For attachments with a color format, this uses the access type VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT.

newtype VkDependencyFlagBits

VkDependencyFlagBits - Bitmask specifying how execution and memory dependencies are formed

See Also

VkDependencyFlags

Constructors

VkDependencyFlagBits VkFlags

pattern VK_DEPENDENCY_BY_REGION_BIT :: VkDependencyFlagBits

VK_DEPENDENCY_BY_REGION_BIT specifies that dependencies will be framebuffer-local.

type VkDependencyFlags = VkDependencyFlagBits

VkDependencyFlags - Bitmask of VkDependencyFlagBits

Description

VkDependencyFlags is a bitmask type for setting a mask of zero or more VkDependencyFlagBits.

See Also

VkDependencyFlagBits, VkSubpassDependency, VkSubpassDependency2KHR, vkCmdPipelineBarrier

type VkFramebuffer = Ptr VkFramebuffer_T

VkFramebuffer - Opaque handle to a framebuffer object

See Also

VkCommandBufferInheritanceInfo, VkRenderPassBeginInfo, vkCreateFramebuffer, vkDestroyFramebuffer

newtype VkFramebufferCreateFlags

VkFramebufferCreateFlags - Reserved for future use

Description

VkFramebufferCreateFlags is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkFramebufferCreateInfo

Constructors

VkFramebufferCreateFlags VkFlags

data VkFramebufferCreateInfo

VkFramebufferCreateInfo - Structure specifying parameters of a newly created framebuffer

Description

Applications must ensure that all accesses to memory that backs image subresources used as attachments in a given renderpass instance either happen-before the load operations for those attachments, or happen-after the store operations for those attachments.

Note

This restriction means that the render pass has full knowledge of all uses of all of the attachments, so that the implementation is able to make correct decisions about when and how to perform layout transitions, when to overlap execution of subpasses, etc.

It is legal for a subpass to use no color or depth/stencil attachments, and rather use shader side effects such as image stores and atomics to produce an output. In this case, the subpass continues to use the width, height, and layers of the framebuffer to define the dimensions of the rendering area, and the rasterizationSamples from each pipeline’s VkPipelineMultisampleStateCreateInfo to define the number of samples used in rasterization; however, if VkPhysicalDeviceFeatures::variableMultisampleRate is VK_FALSE, then all pipelines to be bound with a given zero-attachment subpass must have the same value for VkPipelineMultisampleStateCreateInfo::rasterizationSamples.

Valid Usage

attachmentCount must be equal to the attachment count specified in renderPass

Each element of pAttachments that is used as a color attachment or resolve attachment by renderPass must have been created with a usage value including VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT

Each element of pAttachments that is used as a depth/stencil attachment by renderPass must have been created with a usage value including VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT

Each element of pAttachments that is used as an input attachment by renderPass must have been created with a usage value including VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT

Each element of pAttachments must have been created with an VkFormat value that matches the VkFormat specified by the corresponding VkAttachmentDescription in renderPass

Each element of pAttachments must have been created with a samples value that matches the samples value specified by the corresponding VkAttachmentDescription in renderPass

Each element of pAttachments must have dimensions at least as large as the corresponding framebuffer dimension

Each element of pAttachments must only specify a single mip level

Each element of pAttachments must have been created with the identity swizzle

width must be greater than 0.

width must be less than or equal to VkPhysicalDeviceLimits::maxFramebufferWidth

height must be greater than 0.

height must be less than or equal to VkPhysicalDeviceLimits::maxFramebufferHeight

layers must be greater than 0.

layers must be less than or equal to VkPhysicalDeviceLimits::maxFramebufferLayers

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO

pNext must be NULL

flags must be 0

renderPass must be a valid VkRenderPass handle

If attachmentCount is not 0, pAttachments must be a valid pointer to an array of attachmentCount valid VkImageView handles

Both of renderPass, and the elements of pAttachments that are valid handles must have been created, allocated, or retrieved from the same VkDevice

See Also

VkFramebufferCreateFlags, VkImageView, VkRenderPass, VkStructureType, vkCreateFramebuffer

Constructors

VkFramebufferCreateInfo

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkFlags :: VkFramebufferCreateFlags

flags is reserved for future use.

vkRenderPass :: VkRenderPass

renderPass is a render pass that defines what render passes the framebuffer will be compatible with. See Render Pass Compatibility for details.

vkAttachmentCount :: Word32

attachmentCount is the number of attachments.

vkPAttachments :: Ptr VkImageView

pAttachments is an array of VkImageView handles, each of which will be used as the corresponding attachment in a render pass instance.

vkWidth :: Word32

width, height and layers define the dimensions of the framebuffer.

vkHeight :: Word32

vkLayers :: Word32

newtype VkPipelineBindPoint

VkPipelineBindPoint - Specify the bind point of a pipeline object to a command buffer

See Also

VkDescriptorUpdateTemplateCreateInfo, VkIndirectCommandsLayoutCreateInfoNVX, VkSubpassDescription, VkSubpassDescription2KHR, vkCmdBindDescriptorSets, vkCmdBindPipeline, vkCmdPushDescriptorSetKHR

Constructors

VkPipelineBindPoint Int32

pattern VK_PIPELINE_BIND_POINT_GRAPHICS :: VkPipelineBindPoint

VK_PIPELINE_BIND_POINT_GRAPHICS specifies binding as a graphics pipeline.

pattern VK_PIPELINE_BIND_POINT_COMPUTE :: VkPipelineBindPoint

VK_PIPELINE_BIND_POINT_COMPUTE specifies binding as a compute pipeline.

newtype VkRenderPassCreateFlags

VkRenderPassCreateFlags - Reserved for future use

Description

VkRenderPassCreateFlags is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkRenderPassCreateInfo, VkRenderPassCreateInfo2KHR

Constructors

VkRenderPassCreateFlags VkFlags

data VkRenderPassCreateInfo

VkRenderPassCreateInfo - Structure specifying parameters of a newly created render pass

Description

Note

Care should be taken to avoid a data race here; if any subpasses access attachments with overlapping memory locations, and one of those accesses is a write, a subpass dependency needs to be included between them.

Valid Usage

If the attachment member of any element of pInputAttachments, pColorAttachments, pResolveAttachments or pDepthStencilAttachment, or any element of pPreserveAttachments in any element of pSubpasses is not VK_ATTACHMENT_UNUSED, it must be less than attachmentCount

For any member of pAttachments with a loadOp equal to VK_ATTACHMENT_LOAD_OP_CLEAR, the first use of that attachment must not specify a layout equal to VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL or VK_IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL.

For any member of pAttachments with a stencilLoadOp equal to VK_ATTACHMENT_LOAD_OP_CLEAR, the first use of that attachment must not specify a layout equal to VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL or VK_IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL.

For any element of pDependencies, if the srcSubpass is not VK_SUBPASS_EXTERNAL, all stage flags included in the srcStageMask member of that dependency must be a pipeline stage supported by the pipeline identified by the pipelineBindPoint member of the source subpass

For any element of pDependencies, if the dstSubpass is not VK_SUBPASS_EXTERNAL, all stage flags included in the dstStageMask member of that dependency must be a pipeline stage supported by the pipeline identified by the pipelineBindPoint member of the source subpass

The srcSubpass member of each element of pDependencies must be less than subpassCount

The dstSubpass member of each element of pDependencies must be less than subpassCount

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkRenderPassFragmentDensityMapCreateInfoEXT, VkRenderPassInputAttachmentAspectCreateInfo, or VkRenderPassMultiviewCreateInfo

Each sType member in the pNext chain must be unique

If attachmentCount is not 0, pAttachments must be a valid pointer to an array of attachmentCount valid VkAttachmentDescription structures

pSubpasses must be a valid pointer to an array of subpassCount valid VkSubpassDescription structures

If dependencyCount is not 0, pDependencies must be a valid pointer to an array of dependencyCount valid VkSubpassDependency structures

subpassCount must be greater than 0

See Also

VkAttachmentDescription, VkRenderPassCreateFlags, VkStructureType, VkSubpassDependency, VkSubpassDescription, vkCreateRenderPass

Constructors

VkRenderPassCreateInfo

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkFlags :: VkRenderPassCreateFlags

flags is reserved for future use.

vkAttachmentCount :: Word32

attachmentCount is the number of attachments used by this render pass.

vkPAttachments :: Ptr VkAttachmentDescription

pAttachments points to an array of attachmentCount VkAttachmentDescription structures describing the attachments used by the render pass.

vkSubpassCount :: Word32

subpassCount is the number of subpasses to create.

vkPSubpasses :: Ptr VkSubpassDescription

pSubpasses points to an array of subpassCount VkSubpassDescription structures describing each subpass.

vkDependencyCount :: Word32

dependencyCount is the number of memory dependencies between pairs of subpasses.

vkPDependencies :: Ptr VkSubpassDependency

pDependencies points to an array of dependencyCount VkSubpassDependency structures describing dependencies between pairs of subpasses.

data VkSubpassDependency

VkSubpassDependency - Structure specifying a subpass dependency

Description

If srcSubpass is equal to dstSubpass then the VkSubpassDependency describes a subpass self-dependency, and only constrains the pipeline barriers allowed within a subpass instance. Otherwise, when a render pass instance which includes a subpass dependency is submitted to a queue, it defines a memory dependency between the subpasses identified by srcSubpass and dstSubpass.

If srcSubpass is equal to VK_SUBPASS_EXTERNAL, the first synchronization scope includes commands that occur earlier in submission order than the vkCmdBeginRenderPass used to begin the render pass instance. Otherwise, the first set of commands includes all commands submitted as part of the subpass instance identified by srcSubpass and any load, store or multisample resolve operations on attachments used in srcSubpass. In either case, the first synchronization scope is limited to operations on the pipeline stages determined by the source stage mask specified by srcStageMask.

If dstSubpass is equal to VK_SUBPASS_EXTERNAL, the second synchronization scope includes commands that occur later in submission order than the vkCmdEndRenderPass used to end the render pass instance. Otherwise, the second set of commands includes all commands submitted as part of the subpass instance identified by dstSubpass and any load, store or multisample resolve operations on attachments used in dstSubpass. In either case, the second synchronization scope is limited to operations on the pipeline stages determined by the destination stage mask specified by dstStageMask.

The first access scope is limited to access in the pipeline stages determined by the source stage mask specified by srcStageMask. It is also limited to access types in the source access mask specified by srcAccessMask.

The second access scope is limited to access in the pipeline stages determined by the destination stage mask specified by dstStageMask. It is also limited to access types in the destination access mask specified by dstAccessMask.

The availability and visibility operations defined by a subpass dependency affect the execution of image layout transitions within the render pass.

Note

For non-attachment resources, the memory dependency expressed by subpass dependency is nearly identical to that of a VkMemoryBarrier (with matching srcAccessMask/dstAccessMask parameters) submitted as a part of a vkCmdPipelineBarrier (with matching srcStageMask/dstStageMask parameters). The only difference being that its scopes are limited to the identified subpasses rather than potentially affecting everything before and after.

For attachments however, subpass dependencies work more like a VkImageMemoryBarrier defined similarly to the VkMemoryBarrier above, the queue family indices set to VK_QUEUE_FAMILY_IGNORED, and layouts as follows:

The equivalent to oldLayout is the attachment’s layout according to the subpass description for srcSubpass.

The equivalent to newLayout is the attachment’s layout according to the subpass description for dstSubpass.

Valid Usage

If srcSubpass is not VK_SUBPASS_EXTERNAL, srcStageMask must not include VK_PIPELINE_STAGE_HOST_BIT

If dstSubpass is not VK_SUBPASS_EXTERNAL, dstStageMask must not include VK_PIPELINE_STAGE_HOST_BIT

If the geometry shaders feature is not enabled, srcStageMask must not contain VK_PIPELINE_STAGE_GEOMETRY_SHADER_BIT

If the geometry shaders feature is not enabled, dstStageMask must not contain VK_PIPELINE_STAGE_GEOMETRY_SHADER_BIT

If the tessellation shaders feature is not enabled, srcStageMask must not contain VK_PIPELINE_STAGE_TESSELLATION_CONTROL_SHADER_BIT or VK_PIPELINE_STAGE_TESSELLATION_EVALUATION_SHADER_BIT

If the tessellation shaders feature is not enabled, dstStageMask must not contain VK_PIPELINE_STAGE_TESSELLATION_CONTROL_SHADER_BIT or VK_PIPELINE_STAGE_TESSELLATION_EVALUATION_SHADER_BIT

srcSubpass must be less than or equal to dstSubpass, unless one of them is VK_SUBPASS_EXTERNAL, to avoid cyclic dependencies and ensure a valid execution order

srcSubpass and dstSubpass must not both be equal to VK_SUBPASS_EXTERNAL

If srcSubpass is equal to dstSubpass, srcStageMask and dstStageMask must not set any bits that are neither VK_PIPELINE_STAGE_ALL_GRAPHICS_BIT, nor one of the graphics pipeline stages

If srcSubpass is equal to dstSubpass and not all of the stages in srcStageMask and dstStageMask are framebuffer-space stages, the logically latest pipeline stage in srcStageMask must be logically earlier than or equal to the logically earliest pipeline stage in dstStageMask

Any access flag included in srcAccessMask must be supported by one of the pipeline stages in srcStageMask, as specified in the table of supported access types

Any access flag included in dstAccessMask must be supported by one of the pipeline stages in dstStageMask, as specified in the table of supported access types

If srcSubpass equals dstSubpass, and srcStageMask and dstStageMask both include a framebuffer-space stage, then dependencyFlags must include VK_DEPENDENCY_BY_REGION_BIT

Valid Usage (Implicit)

srcStageMask must be a valid combination of VkPipelineStageFlagBits values

srcStageMask must not be 0

dstStageMask must be a valid combination of VkPipelineStageFlagBits values

dstStageMask must not be 0

srcAccessMask must be a valid combination of VkAccessFlagBits values

dstAccessMask must be a valid combination of VkAccessFlagBits values

dependencyFlags must be a valid combination of VkDependencyFlagBits values

See Also

VkAccessFlags, VkDependencyFlags, VkPipelineStageFlags, VkRenderPassCreateInfo

Constructors

VkSubpassDependency

Fields

vkSrcSubpass :: Word32

srcSubpass is the subpass index of the first subpass in the dependency, or VK_SUBPASS_EXTERNAL.

vkDstSubpass :: Word32

dstSubpass is the subpass index of the second subpass in the dependency, or VK_SUBPASS_EXTERNAL.

vkSrcStageMask :: VkPipelineStageFlags

srcStageMask is a bitmask of VkPipelineStageFlagBits specifying the source stage mask.

vkDstStageMask :: VkPipelineStageFlags

dstStageMask is a bitmask of VkPipelineStageFlagBits specifying the destination stage mask

vkSrcAccessMask :: VkAccessFlags

srcAccessMask is a bitmask of VkAccessFlagBits specifying a source access mask.

vkDstAccessMask :: VkAccessFlags

dstAccessMask is a bitmask of VkAccessFlagBits specifying a destination access mask.

vkDependencyFlags :: VkDependencyFlags

dependencyFlags is a bitmask of VkDependencyFlagBits.

data VkSubpassDescription

VkSubpassDescription - Structure specifying a subpass description

Description

Each element of the pInputAttachments array corresponds to an input attachment index in a fragment shader, i.e. if a shader declares an image variable decorated with a InputAttachmentIndex value of X, then it uses the attachment provided in pInputAttachments[X]. Input attachments must also be bound to the pipeline in a descriptor set. If the attachment member of any element of pInputAttachments is VK_ATTACHMENT_UNUSED, the application must not read from the corresponding input attachment index. Fragment shaders can use subpass input variables to access the contents of an input attachment at the fragment’s (x, y, layer) framebuffer coordinates.

Each element of the pColorAttachments array corresponds to an output location in the shader, i.e. if the shader declares an output variable decorated with a Location value of X, then it uses the attachment provided in pColorAttachments[X]. If the attachment member of any element of pColorAttachments is VK_ATTACHMENT_UNUSED, writes to the corresponding location by a fragment are discarded.

If pResolveAttachments is not NULL, each of its elements corresponds to a color attachment (the element in pColorAttachments at the same index), and a multisample resolve operation is defined for each attachment. At the end of each subpass, multisample resolve operations read the subpass’s color attachments, and resolve the samples for each pixel to the same pixel location in the corresponding resolve attachments, unless the resolve attachment index is VK_ATTACHMENT_UNUSED.

If pDepthStencilAttachment is NULL, or if its attachment index is VK_ATTACHMENT_UNUSED, it indicates that no depth/stencil attachment will be used in the subpass.

The contents of an attachment within the render area become undefined at the start of a subpass S if all of the following conditions are true:

The attachment is used as a color, depth/stencil, or resolve attachment in any subpass in the render pass.

There is a subpass S1 that uses or preserves the attachment, and a subpass dependency from S1 to S.

The attachment is not used or preserved in subpass S.

Once the contents of an attachment become undefined in subpass S, they remain undefined for subpasses in subpass dependency chains starting with subpass S until they are written again. However, they remain valid for subpasses in other subpass dependency chains starting with subpass S1 if those subpasses use or preserve the attachment.

Valid Usage

pipelineBindPoint must be VK_PIPELINE_BIND_POINT_GRAPHICS

colorAttachmentCount must be less than or equal to VkPhysicalDeviceLimits::maxColorAttachments

If the first use of an attachment in this render pass is as an input attachment, and the attachment is not also used as a color or depth/stencil attachment in the same subpass, then loadOp must not be VK_ATTACHMENT_LOAD_OP_CLEAR

If pResolveAttachments is not NULL, for each resolve attachment that is not VK_ATTACHMENT_UNUSED, the corresponding color attachment must not be VK_ATTACHMENT_UNUSED

If pResolveAttachments is not NULL, for each resolve attachment that is not VK_ATTACHMENT_UNUSED, the corresponding color attachment must not have a sample count of VK_SAMPLE_COUNT_1_BIT

If pResolveAttachments is not NULL, each resolve attachment that is not VK_ATTACHMENT_UNUSED must have a sample count of VK_SAMPLE_COUNT_1_BIT

If pResolveAttachments is not NULL, each resolve attachment that is not VK_ATTACHMENT_UNUSED must have the same VkFormat as its corresponding color attachment

All attachments in pColorAttachments that are not VK_ATTACHMENT_UNUSED must have the same sample count

All attachments in pInputAttachments that are not VK_ATTACHMENT_UNUSED must have formats whose features contain at least one of VK_FORMAT_FEATURE_COLOR_ATTACHMENT_BIT or VK_FORMAT_FEATURE_DEPTH_STENCIL_ATTACHMENT_BIT.

All attachments in pColorAttachments that are not VK_ATTACHMENT_UNUSED must have formats whose features contain VK_FORMAT_FEATURE_COLOR_ATTACHMENT_BIT

All attachments in pResolveAttachments that are not VK_ATTACHMENT_UNUSED must have formats whose features contain VK_FORMAT_FEATURE_COLOR_ATTACHMENT_BIT

If pDepthStencilAttachment is not NULL and the attachment is not VK_ATTACHMENT_UNUSED then it must have a format whose features contain VK_FORMAT_FEATURE_DEPTH_STENCIL_ATTACHMENT_BIT

If neither the VK_AMD_mixed_attachment_samples nor the VK_NV_framebuffer_mixed_samples extensions are enabled, and if pDepthStencilAttachment is not VK_ATTACHMENT_UNUSED and any attachments in pColorAttachments are not VK_ATTACHMENT_UNUSED, they must have the same sample count

The attachment member of each element of pPreserveAttachments must not be VK_ATTACHMENT_UNUSED

Each element of pPreserveAttachments must not also be an element of any other member of the subpass description

If any attachment is used by more than one VkAttachmentReference member, then each use must use the same layout

Valid Usage (Implicit)

flags must be a valid combination of VkSubpassDescriptionFlagBits values

pipelineBindPoint must be a valid VkPipelineBindPoint value

If inputAttachmentCount is not 0, pInputAttachments must be a valid pointer to an array of inputAttachmentCount valid VkAttachmentReference structures

If colorAttachmentCount is not 0, pColorAttachments must be a valid pointer to an array of colorAttachmentCount valid VkAttachmentReference structures

If colorAttachmentCount is not 0, and pResolveAttachments is not NULL, pResolveAttachments must be a valid pointer to an array of colorAttachmentCount valid VkAttachmentReference structures

If pDepthStencilAttachment is not NULL, pDepthStencilAttachment must be a valid pointer to a valid VkAttachmentReference structure

If preserveAttachmentCount is not 0, pPreserveAttachments must be a valid pointer to an array of preserveAttachmentCount uint32_t values

See Also

VkAttachmentReference, VkPipelineBindPoint, VkRenderPassCreateInfo, VkSubpassDescriptionFlags

Constructors

VkSubpassDescription

Fields

vkFlags :: VkSubpassDescriptionFlags

flags is a bitmask of VkSubpassDescriptionFlagBits specifying usage of the subpass.

vkPipelineBindPoint :: VkPipelineBindPoint

pipelineBindPoint is a VkPipelineBindPoint value specifying the pipeline type supported for this subpass.

vkInputAttachmentCount :: Word32

inputAttachmentCount is the number of input attachments.

vkPInputAttachments :: Ptr VkAttachmentReference

pInputAttachments is an array of VkAttachmentReference structures defining the input attachments for this subpass and their layouts.

vkColorAttachmentCount :: Word32

colorAttachmentCount is the number of color attachments.

vkPColorAttachments :: Ptr VkAttachmentReference

pColorAttachments is an array of VkAttachmentReference structures defining the color attachments for this subpass and their layouts.

vkPResolveAttachments :: Ptr VkAttachmentReference

pResolveAttachments is an optional array of colorAttachmentCount VkAttachmentReference structures defining the resolve attachments for this subpass and their layouts.

vkPDepthStencilAttachment :: Ptr VkAttachmentReference

pDepthStencilAttachment is a pointer to a VkAttachmentReference specifying the depth/stencil attachment for this subpass and its layout.

vkPreserveAttachmentCount :: Word32

preserveAttachmentCount is the number of preserved attachments.

vkPPreserveAttachments :: Ptr Word32

pPreserveAttachments is an array of preserveAttachmentCount render pass attachment indices identifying attachments that are not used by this subpass, but whose contents must be preserved throughout the subpass.

newtype VkSubpassDescriptionFlagBits

VkSubpassDescriptionFlagBits - Bitmask specifying usage of a subpass

Description

Note

All bits for this type are defined by extensions, and none of those extensions are enabled in this build of the specification.

See Also

VkSubpassDescriptionFlags

Constructors

VkSubpassDescriptionFlagBits VkFlags

type VkSubpassDescriptionFlags = VkSubpassDescriptionFlagBits

VkSubpassDescriptionFlags - Bitmask of VkSubpassDescriptionFlagBits

Description

VkSubpassDescriptionFlags is a bitmask type for setting a mask of zero or more VkSubpassDescriptionFlagBits.

See Also

VkSubpassDescription, VkSubpassDescription2KHR, VkSubpassDescriptionFlagBits

type FN_vkCreateFramebuffer = ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkFramebufferCreateInfo) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pFramebuffer" ::: Ptr VkFramebuffer) -> IO VkResult

type PFN_vkCreateFramebuffer = FunPtr FN_vkCreateFramebuffer

vkCreateFramebuffer :: DeviceCmds -> ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkFramebufferCreateInfo) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pFramebuffer" ::: Ptr VkFramebuffer) -> IO VkResult

vkCreateFramebuffer - Create a new framebuffer object

Parameters

device is the logical device that creates the framebuffer.

pCreateInfo points to a VkFramebufferCreateInfo structure which describes additional information about framebuffer creation.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pFramebuffer points to a VkFramebuffer handle in which the resulting framebuffer object is returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkFramebufferCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pFramebuffer must be a valid pointer to a VkFramebuffer handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDevice, VkFramebuffer, VkFramebufferCreateInfo

type FN_vkCreateRenderPass = ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkRenderPassCreateInfo) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pRenderPass" ::: Ptr VkRenderPass) -> IO VkResult

type PFN_vkCreateRenderPass = FunPtr FN_vkCreateRenderPass

vkCreateRenderPass :: DeviceCmds -> ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkRenderPassCreateInfo) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pRenderPass" ::: Ptr VkRenderPass) -> IO VkResult

vkCreateRenderPass - Create a new render pass object

Parameters

device is the logical device that creates the render pass.

pCreateInfo is a pointer to an instance of the VkRenderPassCreateInfo structure that describes the parameters of the render pass.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pRenderPass points to a VkRenderPass handle in which the resulting render pass object is returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkRenderPassCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pRenderPass must be a valid pointer to a VkRenderPass handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDevice, VkRenderPass, VkRenderPassCreateInfo

type FN_vkDestroyFramebuffer = ("device" ::: VkDevice) -> ("framebuffer" ::: VkFramebuffer) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> IO ()

type PFN_vkDestroyFramebuffer = FunPtr FN_vkDestroyFramebuffer

vkDestroyFramebuffer :: DeviceCmds -> ("device" ::: VkDevice) -> ("framebuffer" ::: VkFramebuffer) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> IO ()

vkDestroyFramebuffer - Destroy a framebuffer object

Parameters

device is the logical device that destroys the framebuffer.

framebuffer is the handle of the framebuffer to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to framebuffer must have completed execution

If VkAllocationCallbacks were provided when framebuffer was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when framebuffer was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid VkDevice handle

If framebuffer is not VK_NULL_HANDLE, framebuffer must be a valid VkFramebuffer handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

If framebuffer is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to framebuffer must be externally synchronized

See Also

VkAllocationCallbacks, VkDevice, VkFramebuffer

type FN_vkDestroyRenderPass = ("device" ::: VkDevice) -> ("renderPass" ::: VkRenderPass) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> IO ()

type PFN_vkDestroyRenderPass = FunPtr FN_vkDestroyRenderPass

vkDestroyRenderPass :: DeviceCmds -> ("device" ::: VkDevice) -> ("renderPass" ::: VkRenderPass) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> IO ()

vkDestroyRenderPass - Destroy a render pass object

Parameters

device is the logical device that destroys the render pass.

renderPass is the handle of the render pass to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to renderPass must have completed execution

If VkAllocationCallbacks were provided when renderPass was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when renderPass was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid VkDevice handle

If renderPass is not VK_NULL_HANDLE, renderPass must be a valid VkRenderPass handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

If renderPass is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to renderPass must be externally synchronized

See Also

VkAllocationCallbacks, VkDevice, VkRenderPass

type FN_vkGetRenderAreaGranularity = ("device" ::: VkDevice) -> ("renderPass" ::: VkRenderPass) -> ("pGranularity" ::: Ptr VkExtent2D) -> IO ()

type PFN_vkGetRenderAreaGranularity = FunPtr FN_vkGetRenderAreaGranularity

vkGetRenderAreaGranularity :: DeviceCmds -> ("device" ::: VkDevice) -> ("renderPass" ::: VkRenderPass) -> ("pGranularity" ::: Ptr VkExtent2D) -> IO ()

vkGetRenderAreaGranularity - Returns the granularity for optimal render area

Parameters

device is the logical device that owns the render pass.

renderPass is a handle to a render pass.

pGranularity points to a VkExtent2D structure in which the granularity is returned.

Description

The conditions leading to an optimal renderArea are:

the offset.x member in renderArea is a multiple of the width member of the returned VkExtent2D (the horizontal granularity).

the offset.y member in renderArea is a multiple of the height of the returned VkExtent2D (the vertical granularity).

either the offset.width member in renderArea is a multiple of the horizontal granularity or offset.x+offset.width is equal to the width of the framebuffer in the VkRenderPassBeginInfo.

either the offset.height member in renderArea is a multiple of the vertical granularity or offset.y+offset.height is equal to the height of the framebuffer in the VkRenderPassBeginInfo.

Subpass dependencies are not affected by the render area, and apply to the entire image subresources attached to the framebuffer as specified in the description of automatic layout transitions. Similarly, pipeline barriers are valid even if their effect extends outside the render area.

Valid Usage (Implicit)

device must be a valid VkDevice handle

renderPass must be a valid VkRenderPass handle

pGranularity must be a valid pointer to a VkExtent2D structure

renderPass must have been created, allocated, or retrieved from device

See Also

VkDevice, VkExtent2D, VkRenderPass