Graphics.Vulkan.C.Core10.PipelineLayout

Documentation

type VkDescriptorSetLayout = Ptr VkDescriptorSetLayout_T

VkDescriptorSetLayout - Opaque handle to a descriptor set layout object

See Also

VkDescriptorSetAllocateInfo, VkDescriptorUpdateTemplateCreateInfo, VkPipelineLayoutCreateInfo, vkCreateDescriptorSetLayout, vkDestroyDescriptorSetLayout

newtype VkPipelineLayoutCreateFlags

VkPipelineLayoutCreateFlags - Reserved for future use

Description

VkPipelineLayoutCreateFlags is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkPipelineLayoutCreateInfo

Constructors

VkPipelineLayoutCreateFlags VkFlags

data VkPipelineLayoutCreateInfo

VkPipelineLayoutCreateInfo - Structure specifying the parameters of a newly created pipeline layout object

Valid Usage

setLayoutCount must be less than or equal to VkPhysicalDeviceLimits::maxBoundDescriptorSets

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_SAMPLER and VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER accessible to any shader stage across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxPerStageDescriptorSamplers

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER and VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC accessible to any shader stage across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxPerStageDescriptorUniformBuffers

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_STORAGE_BUFFER and VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC accessible to any shader stage across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxPerStageDescriptorStorageBuffers

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE, and VK_DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER accessible to any shader stage across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxPerStageDescriptorSampledImages

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, and VK_DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER accessible to any shader stage across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxPerStageDescriptorStorageImages

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT accessible to any given shader stage across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxPerStageDescriptorInputAttachments

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_SAMPLER and VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxDescriptorSetSamplers

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxDescriptorSetUniformBuffers

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxDescriptorSetUniformBuffersDynamic

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_STORAGE_BUFFER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxDescriptorSetStorageBuffers

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxDescriptorSetStorageBuffersDynamic

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE, and VK_DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxDescriptorSetSampledImages

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, and VK_DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxDescriptorSetStorageImages

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxDescriptorSetInputAttachments

Any two elements of pPushConstantRanges must not include the same stage in stageFlags

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO

pNext must be NULL

flags must be 0

If setLayoutCount is not 0, pSetLayouts must be a valid pointer to an array of setLayoutCount valid VkDescriptorSetLayout handles

If pushConstantRangeCount is not 0, pPushConstantRanges must be a valid pointer to an array of pushConstantRangeCount valid VkPushConstantRange structures

See Also

VkDescriptorSetLayout, VkPipelineLayoutCreateFlags, VkPushConstantRange, VkStructureType, vkCreatePipelineLayout

Constructors

VkPipelineLayoutCreateInfo

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkFlags :: VkPipelineLayoutCreateFlags

flags is reserved for future use.

vkSetLayoutCount :: Word32

setLayoutCount is the number of descriptor sets included in the pipeline layout.

vkPSetLayouts :: Ptr VkDescriptorSetLayout

pSetLayouts is a pointer to an array of VkDescriptorSetLayout objects.

vkPushConstantRangeCount :: Word32

pushConstantRangeCount is the number of push constant ranges included in the pipeline layout.

vkPPushConstantRanges :: Ptr VkPushConstantRange

pPushConstantRanges is a pointer to an array of VkPushConstantRange structures defining a set of push constant ranges for use in a single pipeline layout. In addition to descriptor set layouts, a pipeline layout also describes how many push constants can be accessed by each stage of the pipeline.

Note

Push constants represent a high speed path to modify constant data in pipelines that is expected to outperform memory-backed resource updates.

data VkPushConstantRange

VkPushConstantRange - Structure specifying a push constant range

Valid Usage (Implicit)

See Also

VkPipelineLayoutCreateInfo, VkShaderStageFlags

Constructors

VkPushConstantRange

Fields

vkStageFlags :: VkShaderStageFlags

stageFlags must not be 0

vkOffset :: Word32

offset must be a multiple of 4

vkSize :: Word32

size must be less than or equal to VkPhysicalDeviceLimits::maxPushConstantsSize minus offset

type VkShaderStageFlags = VkShaderStageFlagBits

VkShaderStageFlags - Bitmask of VkShaderStageFlagBits

Description

VkShaderStageFlags is a bitmask type for setting a mask of zero or more VkShaderStageFlagBits.

See Also

VkDescriptorSetLayoutBinding, VkObjectTablePushConstantEntryNVX, VkPhysicalDeviceCooperativeMatrixPropertiesNV, VkPhysicalDeviceSubgroupProperties, VkPushConstantRange, VkShaderStageFlagBits, VkShaderStatisticsInfoAMD, vkCmdPushConstants

type FN_vkCreatePipelineLayout = ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkPipelineLayoutCreateInfo) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pPipelineLayout" ::: Ptr VkPipelineLayout) -> IO VkResult

type PFN_vkCreatePipelineLayout = FunPtr FN_vkCreatePipelineLayout

vkCreatePipelineLayout :: DeviceCmds -> ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkPipelineLayoutCreateInfo) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pPipelineLayout" ::: Ptr VkPipelineLayout) -> IO VkResult

vkCreatePipelineLayout - Creates a new pipeline layout object

Parameters

device is the logical device that creates the pipeline layout.

pCreateInfo is a pointer to an instance of the VkPipelineLayoutCreateInfo structure specifying the state of the pipeline layout object.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pPipelineLayout points to a VkPipelineLayout handle in which the resulting pipeline layout object is returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkPipelineLayoutCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pPipelineLayout must be a valid pointer to a VkPipelineLayout handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDevice, VkPipelineLayout, VkPipelineLayoutCreateInfo

type FN_vkDestroyPipelineLayout = ("device" ::: VkDevice) -> ("pipelineLayout" ::: VkPipelineLayout) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> IO ()

type PFN_vkDestroyPipelineLayout = FunPtr FN_vkDestroyPipelineLayout

vkDestroyPipelineLayout :: DeviceCmds -> ("device" ::: VkDevice) -> ("pipelineLayout" ::: VkPipelineLayout) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> IO ()

vkDestroyPipelineLayout - Destroy a pipeline layout object

Parameters

device is the logical device that destroys the pipeline layout.

pipelineLayout is the pipeline layout to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

If VkAllocationCallbacks were provided when pipelineLayout was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when pipelineLayout was created, pAllocator must be NULL

pipelineLayout must not have been passed to any vkCmd* command for any command buffers that are still in the recording state when vkDestroyPipelineLayout is called

Valid Usage (Implicit)

device must be a valid VkDevice handle

If pipelineLayout is not VK_NULL_HANDLE, pipelineLayout must be a valid VkPipelineLayout handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

If pipelineLayout is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to pipelineLayout must be externally synchronized

See Also

VkAllocationCallbacks, VkDevice, VkPipelineLayout