Graphics.Vulkan.C.Core10.QueueSemaphore

Documentation

newtype VkSemaphoreCreateFlags

VkSemaphoreCreateFlags - Reserved for future use

Description

VkSemaphoreCreateFlags is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkSemaphoreCreateInfo

Constructors

VkSemaphoreCreateFlags VkFlags

data VkSemaphoreCreateInfo

VkSemaphoreCreateInfo - Structure specifying parameters of a newly created semaphore

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_SEMAPHORE_CREATE_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkExportSemaphoreCreateInfo or VkExportSemaphoreWin32HandleInfoKHR

Each sType member in the pNext chain must be unique

flags must be 0

See Also

VkSemaphoreCreateFlags, VkStructureType, vkCreateSemaphore

Constructors

VkSemaphoreCreateInfo

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkFlags :: VkSemaphoreCreateFlags

flags is reserved for future use.

type FN_vkCreateSemaphore = ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkSemaphoreCreateInfo) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pSemaphore" ::: Ptr VkSemaphore) -> IO VkResult

type PFN_vkCreateSemaphore = FunPtr FN_vkCreateSemaphore

vkCreateSemaphore :: DeviceCmds -> ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkSemaphoreCreateInfo) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pSemaphore" ::: Ptr VkSemaphore) -> IO VkResult

vkCreateSemaphore - Create a new queue semaphore object

Parameters

device is the logical device that creates the semaphore.

pCreateInfo is a pointer to an instance of the VkSemaphoreCreateInfo structure which contains information about how the semaphore is to be created.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pSemaphore points to a handle in which the resulting semaphore object is returned.

Description

When created, the semaphore is in the unsignaled state.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkSemaphoreCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSemaphore must be a valid pointer to a VkSemaphore handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDevice, VkSemaphore, VkSemaphoreCreateInfo

type FN_vkDestroySemaphore = ("device" ::: VkDevice) -> ("semaphore" ::: VkSemaphore) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> IO ()

type PFN_vkDestroySemaphore = FunPtr FN_vkDestroySemaphore

vkDestroySemaphore :: DeviceCmds -> ("device" ::: VkDevice) -> ("semaphore" ::: VkSemaphore) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> IO ()

vkDestroySemaphore - Destroy a semaphore object

Parameters

device is the logical device that destroys the semaphore.

semaphore is the handle of the semaphore to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted batches that refer to semaphore must have completed execution

If VkAllocationCallbacks were provided when semaphore was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when semaphore was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid VkDevice handle

If semaphore is not VK_NULL_HANDLE, semaphore must be a valid VkSemaphore handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

If semaphore is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to semaphore must be externally synchronized

See Also

VkAllocationCallbacks, VkDevice, VkSemaphore