Graphics.Vulkan.C.Core10.SparseResourceMemoryManagement

Documentation

data VkBindSparseInfo

VkBindSparseInfo - Structure specifying a sparse binding operation

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_BIND_SPARSE_INFO

pNext must be NULL or a pointer to a valid instance of VkDeviceGroupBindSparseInfo

If waitSemaphoreCount is not 0, pWaitSemaphores must be a valid pointer to an array of waitSemaphoreCount valid VkSemaphore handles

If bufferBindCount is not 0, pBufferBinds must be a valid pointer to an array of bufferBindCount valid VkSparseBufferMemoryBindInfo structures

If imageOpaqueBindCount is not 0, pImageOpaqueBinds must be a valid pointer to an array of imageOpaqueBindCount valid VkSparseImageOpaqueMemoryBindInfo structures

If imageBindCount is not 0, pImageBinds must be a valid pointer to an array of imageBindCount valid VkSparseImageMemoryBindInfo structures

If signalSemaphoreCount is not 0, pSignalSemaphores must be a valid pointer to an array of signalSemaphoreCount valid VkSemaphore handles

Both of the elements of pSignalSemaphores, and the elements of pWaitSemaphores that are valid handles must have been created, allocated, or retrieved from the same VkDevice

See Also

VkSemaphore, VkSparseBufferMemoryBindInfo, VkSparseImageMemoryBindInfo, VkSparseImageOpaqueMemoryBindInfo, VkStructureType, vkQueueBindSparse

Constructors

VkBindSparseInfo

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkWaitSemaphoreCount :: Word32

waitSemaphoreCount is the number of semaphores upon which to wait before executing the sparse binding operations for the batch.

vkPWaitSemaphores :: Ptr VkSemaphore

pWaitSemaphores is a pointer to an array of semaphores upon which to wait on before the sparse binding operations for this batch begin execution. If semaphores to wait on are provided, they define a semaphore wait operation.

vkBufferBindCount :: Word32

bufferBindCount is the number of sparse buffer bindings to perform in the batch.

vkPBufferBinds :: Ptr VkSparseBufferMemoryBindInfo

pBufferBinds is a pointer to an array of VkSparseBufferMemoryBindInfo structures.

vkImageOpaqueBindCount :: Word32

imageOpaqueBindCount is the number of opaque sparse image bindings to perform.

vkPImageOpaqueBinds :: Ptr VkSparseImageOpaqueMemoryBindInfo

pImageOpaqueBinds is a pointer to an array of VkSparseImageOpaqueMemoryBindInfo structures, indicating opaque sparse image bindings to perform.

vkImageBindCount :: Word32

imageBindCount is the number of sparse image bindings to perform.

vkPImageBinds :: Ptr VkSparseImageMemoryBindInfo

pImageBinds is a pointer to an array of VkSparseImageMemoryBindInfo structures, indicating sparse image bindings to perform.

vkSignalSemaphoreCount :: Word32

signalSemaphoreCount is the number of semaphores to be signaled once the sparse binding operations specified by the structure have completed execution.

vkPSignalSemaphores :: Ptr VkSemaphore

pSignalSemaphores is a pointer to an array of semaphores which will be signaled when the sparse binding operations for this batch have completed execution. If semaphores to be signaled are provided, they define a semaphore signal operation.

newtype VkImageAspectFlagBits

VkImageAspectFlagBits - Bitmask specifying which aspects of an image are included in a view

See Also

VkBindImagePlaneMemoryInfo, VkImageAspectFlags, VkImagePlaneMemoryRequirementsInfo

Constructors

VkImageAspectFlagBits VkFlags

pattern VK_IMAGE_ASPECT_COLOR_BIT :: VkImageAspectFlagBits

VK_IMAGE_ASPECT_COLOR_BIT specifies the color aspect.

pattern VK_IMAGE_ASPECT_DEPTH_BIT :: VkImageAspectFlagBits

VK_IMAGE_ASPECT_DEPTH_BIT specifies the depth aspect.

pattern VK_IMAGE_ASPECT_STENCIL_BIT :: VkImageAspectFlagBits

VK_IMAGE_ASPECT_STENCIL_BIT specifies the stencil aspect.

pattern VK_IMAGE_ASPECT_METADATA_BIT :: VkImageAspectFlagBits

VK_IMAGE_ASPECT_METADATA_BIT specifies the metadata aspect, used for sparse sparse resource operations.

type VkImageAspectFlags = VkImageAspectFlagBits

VkImageAspectFlags - Bitmask of VkImageAspectFlagBits

Description

VkImageAspectFlags is a bitmask type for setting a mask of zero or more VkImageAspectFlagBits.

See Also

VkAttachmentReference2KHR, VkClearAttachment, VkImageAspectFlagBits, VkImageSubresource, VkImageSubresourceLayers, VkImageSubresourceRange, VkInputAttachmentAspectReference, VkSparseImageFormatProperties

data VkImageSubresource

VkImageSubresource - Structure specifying an image subresource

Valid Usage (Implicit)

See Also

VkImageAspectFlags, VkSparseImageMemoryBind, vkGetImageSubresourceLayout

Constructors

VkImageSubresource

Fields

vkAspectMask :: VkImageAspectFlags

aspectMask must not be 0

vkMipLevel :: Word32

mipLevel selects the mipmap level.

vkArrayLayer :: Word32

arrayLayer selects the array layer.

data VkOffset3D

VkOffset3D - Structure specifying a three-dimensional offset

See Also

VkBufferImageCopy, VkImageBlit, VkImageCopy, VkImageResolve, VkSparseImageMemoryBind

Constructors

VkOffset3D

Fields

vkX :: Int32

x is the x offset.

vkY :: Int32

y is the y offset.

vkZ :: Int32

z is the z offset.

data VkSparseBufferMemoryBindInfo

VkSparseBufferMemoryBindInfo - Structure specifying a sparse buffer memory bind operation

Valid Usage (Implicit)

See Also

VkBindSparseInfo, VkBuffer, VkSparseMemoryBind

Constructors

VkSparseBufferMemoryBindInfo

Fields

vkBuffer :: VkBuffer

buffer must be a valid VkBuffer handle

vkBindCount :: Word32

bindCount must be greater than 0

vkPBinds :: Ptr VkSparseMemoryBind

pBinds must be a valid pointer to an array of bindCount valid VkSparseMemoryBind structures

newtype VkSparseImageFormatFlagBits

VkSparseImageFormatFlagBits - Bitmask specifying additional information about a sparse image resource

See Also

VkSparseImageFormatFlags

Constructors

VkSparseImageFormatFlagBits VkFlags

pattern VK_SPARSE_IMAGE_FORMAT_SINGLE_MIPTAIL_BIT :: VkSparseImageFormatFlagBits

VK_SPARSE_IMAGE_FORMAT_SINGLE_MIPTAIL_BIT specifies that the image uses a single mip tail region for all array layers.

pattern VK_SPARSE_IMAGE_FORMAT_ALIGNED_MIP_SIZE_BIT :: VkSparseImageFormatFlagBits

VK_SPARSE_IMAGE_FORMAT_ALIGNED_MIP_SIZE_BIT specifies that the first mip level whose dimensions are not integer multiples of the corresponding dimensions of the sparse image block begins the mip tail region.

pattern VK_SPARSE_IMAGE_FORMAT_NONSTANDARD_BLOCK_SIZE_BIT :: VkSparseImageFormatFlagBits

VK_SPARSE_IMAGE_FORMAT_NONSTANDARD_BLOCK_SIZE_BIT specifies that the image uses non-standard sparse image block dimensions, and the imageGranularity values do not match the standard sparse image block dimensions for the given format.

type VkSparseImageFormatFlags = VkSparseImageFormatFlagBits

VkSparseImageFormatFlags - Bitmask of VkSparseImageFormatFlagBits

Description

VkSparseImageFormatFlags is a bitmask type for setting a mask of zero or more VkSparseImageFormatFlagBits.

See Also

VkSparseImageFormatFlagBits, VkSparseImageFormatProperties

data VkSparseImageFormatProperties

VkSparseImageFormatProperties - Structure specifying sparse image format properties

See Also

VkExtent3D, VkImageAspectFlags, VkSparseImageFormatFlags, VkSparseImageFormatProperties2, VkSparseImageMemoryRequirements, vkGetPhysicalDeviceSparseImageFormatProperties

Constructors

VkSparseImageFormatProperties

Fields

vkAspectMask :: VkImageAspectFlags

aspectMask is a bitmask VkImageAspectFlagBits specifying which aspects of the image the properties apply to.

vkImageGranularity :: VkExtent3D

imageGranularity is the width, height, and depth of the sparse image block in texels or compressed texel blocks.

vkFlags :: VkSparseImageFormatFlags

flags is a bitmask of VkSparseImageFormatFlagBits specifying additional information about the sparse resource.

data VkSparseImageMemoryBind

VkSparseImageMemoryBind - Structure specifying sparse image memory bind

Valid Usage

If the sparse aliased residency feature is not enabled, and if any other resources are bound to ranges of memory, the range of memory being bound must not overlap with those bound ranges

memory and memoryOffset must match the memory requirements of the calling command’s image, as described in section https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#resources-association

subresource must be a valid image subresource for image (see https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#resources-image-views)

offset.x must be a multiple of the sparse image block width (VkSparseImageFormatProperties::imageGranularity.width) of the image

extent.width must either be a multiple of the sparse image block width of the image, or else (extent.width + offset.x) must equal the width of the image subresource

offset.y must be a multiple of the sparse image block height (VkSparseImageFormatProperties::imageGranularity.height) of the image

extent.height must either be a multiple of the sparse image block height of the image, or else (extent.height + offset.y) must equal the height of the image subresource

offset.z must be a multiple of the sparse image block depth (VkSparseImageFormatProperties::imageGranularity.depth) of the image

extent.depth must either be a multiple of the sparse image block depth of the image, or else (extent.depth + offset.z) must equal the depth of the image subresource

Valid Usage (Implicit)

subresource must be a valid VkImageSubresource structure

If memory is not VK_NULL_HANDLE, memory must be a valid VkDeviceMemory handle

flags must be a valid combination of VkSparseMemoryBindFlagBits values

See Also

VkDeviceMemory, VkDeviceSize, VkExtent3D, VkImageSubresource, VkOffset3D, VkSparseImageMemoryBindInfo, VkSparseMemoryBindFlags

Constructors

VkSparseImageMemoryBind

Fields

vkSubresource :: VkImageSubresource

subresource is the image aspect and region of interest in the image.

vkOffset :: VkOffset3D

offset are the coordinates of the first texel within the image subresource to bind.

vkExtent :: VkExtent3D

extent is the size in texels of the region within the image subresource to bind. The extent must be a multiple of the sparse image block dimensions, except when binding sparse image blocks along the edge of an image subresource it can instead be such that any coordinate of offset + extent equals the corresponding dimensions of the image subresource.

vkMemory :: VkDeviceMemory

memory is the VkDeviceMemory object that the sparse image blocks of the image are bound to. If memory is VK_NULL_HANDLE, the sparse image blocks are unbound.

vkMemoryOffset :: VkDeviceSize

memoryOffset is an offset into VkDeviceMemory object. If memory is VK_NULL_HANDLE, this value is ignored.

vkFlags :: VkSparseMemoryBindFlags

flags are sparse memory binding flags.

data VkSparseImageMemoryBindInfo

VkSparseImageMemoryBindInfo - Structure specifying sparse image memory bind info

Valid Usage

The subresource.mipLevel member of each element of pBinds must be less than the mipLevels specified in VkImageCreateInfo when image was created

The subresource.arrayLayer member of each element of pBinds must be less than the arrayLayers specified in VkImageCreateInfo when image was created

Valid Usage (Implicit)

image must be a valid VkImage handle

pBinds must be a valid pointer to an array of bindCount valid VkSparseImageMemoryBind structures

bindCount must be greater than 0

See Also

VkBindSparseInfo, VkImage, VkSparseImageMemoryBind

Constructors

VkSparseImageMemoryBindInfo

Fields

vkImage :: VkImage

image is the VkImage object to be bound

vkBindCount :: Word32

bindCount is the number of VkSparseImageMemoryBind structures in pBinds array

vkPBinds :: Ptr VkSparseImageMemoryBind

pBinds is a pointer to array of VkSparseImageMemoryBind structures

data VkSparseImageMemoryRequirements

VkSparseImageMemoryRequirements - Structure specifying sparse image memory requirements

See Also

VkDeviceSize, VkSparseImageFormatProperties, VkSparseImageMemoryRequirements2, vkGetImageSparseMemoryRequirements

Constructors

VkSparseImageMemoryRequirements

Fields

vkFormatProperties :: VkSparseImageFormatProperties

vkImageMipTailFirstLod :: Word32

imageMipTailFirstLod is the first mip level at which image subresources are included in the mip tail region.

vkImageMipTailSize :: VkDeviceSize

imageMipTailSize is the memory size (in bytes) of the mip tail region. If formatProperties.flags contains VK_SPARSE_IMAGE_FORMAT_SINGLE_MIPTAIL_BIT, this is the size of the whole mip tail, otherwise this is the size of the mip tail of a single array layer. This value is guaranteed to be a multiple of the sparse block size in bytes.

vkImageMipTailOffset :: VkDeviceSize

imageMipTailOffset is the opaque memory offset used with VkSparseImageOpaqueMemoryBindInfo to bind the mip tail region(s).

vkImageMipTailStride :: VkDeviceSize

imageMipTailStride is the offset stride between each array-layer’s mip tail, if formatProperties.flags does not contain VK_SPARSE_IMAGE_FORMAT_SINGLE_MIPTAIL_BIT (otherwise the value is undefined).

data VkSparseImageOpaqueMemoryBindInfo

VkSparseImageOpaqueMemoryBindInfo - Structure specifying sparse image opaque memory bind info

Valid Usage

If the flags member of any element of pBinds contains VK_SPARSE_MEMORY_BIND_METADATA_BIT, the binding range defined must be within the mip tail region of the metadata aspect of image

Valid Usage (Implicit)

image must be a valid VkImage handle

pBinds must be a valid pointer to an array of bindCount valid VkSparseMemoryBind structures

bindCount must be greater than 0

See Also

VkBindSparseInfo, VkImage, VkSparseMemoryBind

Constructors

VkSparseImageOpaqueMemoryBindInfo

Fields

vkImage :: VkImage

image is the VkImage object to be bound.

vkBindCount :: Word32

bindCount is the number of VkSparseMemoryBind structures in the pBinds array.

vkPBinds :: Ptr VkSparseMemoryBind

pBinds is a pointer to array of VkSparseMemoryBind structures.

data VkSparseMemoryBind

VkSparseMemoryBind - Structure specifying a sparse memory bind operation

Description

The binding range [resourceOffset, resourceOffset + size) has different constraints based on flags. If flags contains VK_SPARSE_MEMORY_BIND_METADATA_BIT, the binding range must be within the mip tail region of the metadata aspect. This metadata region is defined by:

metadataRegion = [base, base + imageMipTailSize)

base = imageMipTailOffset + imageMipTailStride × n

and imageMipTailOffset, imageMipTailSize, and imageMipTailStride values are from the VkSparseImageMemoryRequirements corresponding to the metadata aspect of the image, and n is a valid array layer index for the image,

imageMipTailStride is considered to be zero for aspects where VkSparseImageMemoryRequirements::formatProperties.flags contains VK_SPARSE_IMAGE_FORMAT_SINGLE_MIPTAIL_BIT.

If flags does not contain VK_SPARSE_MEMORY_BIND_METADATA_BIT, the binding range must be within the range [0,VkMemoryRequirements::size).

Valid Usage

If memory is not VK_NULL_HANDLE, memory and memoryOffset must match the memory requirements of the resource, as described in section https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#resources-association

If memory is not VK_NULL_HANDLE, memory must not have been created with a memory type that reports VK_MEMORY_PROPERTY_LAZILY_ALLOCATED_BIT bit set

size must be greater than 0

resourceOffset must be less than the size of the resource

size must be less than or equal to the size of the resource minus resourceOffset

memoryOffset must be less than the size of memory

size must be less than or equal to the size of memory minus memoryOffset

Valid Usage (Implicit)

If memory is not VK_NULL_HANDLE, memory must be a valid VkDeviceMemory handle

flags must be a valid combination of VkSparseMemoryBindFlagBits values

See Also

VkDeviceMemory, VkDeviceSize, VkSparseBufferMemoryBindInfo, VkSparseImageOpaqueMemoryBindInfo, VkSparseMemoryBindFlags

Constructors

VkSparseMemoryBind

Fields

vkResourceOffset :: VkDeviceSize

resourceOffset is the offset into the resource.

vkSize :: VkDeviceSize

size is the size of the memory region to be bound.

vkMemory :: VkDeviceMemory

memory is the VkDeviceMemory object that the range of the resource is bound to. If memory is VK_NULL_HANDLE, the range is unbound.

vkMemoryOffset :: VkDeviceSize

memoryOffset is the offset into the VkDeviceMemory object to bind the resource range to. If memory is VK_NULL_HANDLE, this value is ignored.

vkFlags :: VkSparseMemoryBindFlags

flags is a bitmask of VkSparseMemoryBindFlagBits specifying usage of the binding operation.

newtype VkSparseMemoryBindFlagBits

VkSparseMemoryBindFlagBits - Bitmask specifying usage of a sparse memory binding operation

See Also

VkSparseMemoryBindFlags

Constructors

VkSparseMemoryBindFlagBits VkFlags

pattern VK_SPARSE_MEMORY_BIND_METADATA_BIT :: VkSparseMemoryBindFlagBits

VK_SPARSE_MEMORY_BIND_METADATA_BIT specifies that the memory being bound is only for the metadata aspect.

type VkSparseMemoryBindFlags = VkSparseMemoryBindFlagBits

VkSparseMemoryBindFlags - Bitmask of VkSparseMemoryBindFlagBits

Description

VkSparseMemoryBindFlags is a bitmask type for setting a mask of zero or more VkSparseMemoryBindFlagBits.

See Also

VkSparseImageMemoryBind, VkSparseMemoryBind, VkSparseMemoryBindFlagBits

type FN_vkGetImageSparseMemoryRequirements = ("device" ::: VkDevice) -> ("image" ::: VkImage) -> ("pSparseMemoryRequirementCount" ::: Ptr Word32) -> ("pSparseMemoryRequirements" ::: Ptr VkSparseImageMemoryRequirements) -> IO ()

type PFN_vkGetImageSparseMemoryRequirements = FunPtr FN_vkGetImageSparseMemoryRequirements

vkGetImageSparseMemoryRequirements :: DeviceCmds -> ("device" ::: VkDevice) -> ("image" ::: VkImage) -> ("pSparseMemoryRequirementCount" ::: Ptr Word32) -> ("pSparseMemoryRequirements" ::: Ptr VkSparseImageMemoryRequirements) -> IO ()

vkGetImageSparseMemoryRequirements - Query the memory requirements for a sparse image

Parameters

device is the logical device that owns the image.

image is the VkImage object to get the memory requirements for.

pSparseMemoryRequirementCount is a pointer to an integer related to the number of sparse memory requirements available or queried, as described below.

pSparseMemoryRequirements is either NULL or a pointer to an array of VkSparseImageMemoryRequirements structures.

Description

If pSparseMemoryRequirements is NULL, then the number of sparse memory requirements available is returned in pSparseMemoryRequirementCount. Otherwise, pSparseMemoryRequirementCount must point to a variable set by the user to the number of elements in the pSparseMemoryRequirements array, and on return the variable is overwritten with the number of structures actually written to pSparseMemoryRequirements. If pSparseMemoryRequirementCount is less than the number of sparse memory requirements available, at most pSparseMemoryRequirementCount structures will be written.

If the image was not created with VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT then pSparseMemoryRequirementCount will be set to zero and pSparseMemoryRequirements will not be written to.

Note

It is legal for an implementation to report a larger value in VkMemoryRequirements::size than would be obtained by adding together memory sizes for all VkSparseImageMemoryRequirements returned by vkGetImageSparseMemoryRequirements. This may occur when the implementation requires unused padding in the address range describing the resource.

Valid Usage (Implicit)

device must be a valid VkDevice handle

image must be a valid VkImage handle

pSparseMemoryRequirementCount must be a valid pointer to a uint32_t value

If the value referenced by pSparseMemoryRequirementCount is not 0, and pSparseMemoryRequirements is not NULL, pSparseMemoryRequirements must be a valid pointer to an array of pSparseMemoryRequirementCount VkSparseImageMemoryRequirements structures

image must have been created, allocated, or retrieved from device

See Also

VkDevice, VkImage, VkSparseImageMemoryRequirements

type FN_vkGetPhysicalDeviceSparseImageFormatProperties = ("physicalDevice" ::: VkPhysicalDevice) -> ("format" ::: VkFormat) -> ("type" ::: VkImageType) -> ("samples" ::: VkSampleCountFlagBits) -> ("usage" ::: VkImageUsageFlags) -> ("tiling" ::: VkImageTiling) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkSparseImageFormatProperties) -> IO ()

type PFN_vkGetPhysicalDeviceSparseImageFormatProperties = FunPtr FN_vkGetPhysicalDeviceSparseImageFormatProperties

vkGetPhysicalDeviceSparseImageFormatProperties :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("format" ::: VkFormat) -> ("type" ::: VkImageType) -> ("samples" ::: VkSampleCountFlagBits) -> ("usage" ::: VkImageUsageFlags) -> ("tiling" ::: VkImageTiling) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkSparseImageFormatProperties) -> IO ()

vkGetPhysicalDeviceSparseImageFormatProperties - Retrieve properties of an image format applied to sparse images

Parameters

physicalDevice is the physical device from which to query the sparse image capabilities.

format is the image format.

type is the dimensionality of image.

samples is the number of samples per texel as defined in VkSampleCountFlagBits.

usage is a bitmask describing the intended usage of the image.

tiling is the tiling arrangement of the texel blocks in memory.

pPropertyCount is a pointer to an integer related to the number of sparse format properties available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkSparseImageFormatProperties structures.

Description

If pProperties is NULL, then the number of sparse format properties available is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If pPropertyCount is less than the number of sparse format properties available, at most pPropertyCount structures will be written.

If VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT is not supported for the given arguments, pPropertyCount will be set to zero upon return, and no data will be written to pProperties.

Multiple aspects are returned for depth/stencil images that are implemented as separate planes by the implementation. The depth and stencil data planes each have unique VkSparseImageFormatProperties data.

Depth/stencil images with depth and stencil data interleaved into a single plane will return a single VkSparseImageFormatProperties structure with the aspectMask set to VK_IMAGE_ASPECT_DEPTH_BIT | VK_IMAGE_ASPECT_STENCIL_BIT.

Valid Usage

samples must be a bit value that is set in VkImageFormatProperties::sampleCounts returned by vkGetPhysicalDeviceImageFormatProperties with format, type, tiling, and usage equal to those in this command and flags equal to the value that is set in VkImageCreateInfo::flags when the image is created

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

format must be a valid VkFormat value

type must be a valid VkImageType value

samples must be a valid VkSampleCountFlagBits value

usage must be a valid combination of VkImageUsageFlagBits values

usage must not be 0

tiling must be a valid VkImageTiling value

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkSparseImageFormatProperties structures

See Also

VkFormat, VkImageTiling, VkImageType, VkImageUsageFlags, VkPhysicalDevice, VkSampleCountFlagBits, VkSparseImageFormatProperties

type FN_vkQueueBindSparse = ("queue" ::: VkQueue) -> ("bindInfoCount" ::: Word32) -> ("pBindInfo" ::: Ptr VkBindSparseInfo) -> ("fence" ::: VkFence) -> IO VkResult

type PFN_vkQueueBindSparse = FunPtr FN_vkQueueBindSparse

vkQueueBindSparse :: DeviceCmds -> ("queue" ::: VkQueue) -> ("bindInfoCount" ::: Word32) -> ("pBindInfo" ::: Ptr VkBindSparseInfo) -> ("fence" ::: VkFence) -> IO VkResult

vkQueueBindSparse - Bind device memory to a sparse resource object

Parameters

queue is the queue that the sparse binding operations will be submitted to.

bindInfoCount is the number of elements in the pBindInfo array.

pBindInfo is an array of VkBindSparseInfo structures, each specifying a sparse binding submission batch.

fence is an optional handle to a fence to be signaled. If fence is not VK_NULL_HANDLE, it defines a fence signal operation.

Description

vkQueueBindSparse is a queue submission command, with each batch defined by an element of pBindInfo as an instance of the VkBindSparseInfo structure. Batches begin execution in the order they appear in pBindInfo, but may complete out of order.

Within a batch, a given range of a resource must not be bound more than once. Across batches, if a range is to be bound to one allocation and offset and then to another allocation and offset, then the application must guarantee (usually using semaphores) that the binding operations are executed in the correct order, as well as to order binding operations against the execution of command buffer submissions.

As no operation to vkQueueBindSparse causes any pipeline stage to access memory, synchronization primitives used in this command effectively only define execution dependencies.

Additional information about fence and semaphore operation is described in the synchronization chapter.

Valid Usage

If fence is not VK_NULL_HANDLE, fence must be unsignaled

If fence is not VK_NULL_HANDLE, fence must not be associated with any other queue command that has not yet completed execution on that queue

Each element of the pSignalSemaphores member of each element of pBindInfo must be unsignaled when the semaphore signal operation it defines is executed on the device

When a semaphore unsignal operation defined by any element of the pWaitSemaphores member of any element of pBindInfo executes on queue, no other queue must be waiting on the same semaphore.

All elements of the pWaitSemaphores member of all elements of pBindInfo must be semaphores that are signaled, or have semaphore signal operations previously submitted for execution.

Valid Usage (Implicit)

queue must be a valid VkQueue handle

If bindInfoCount is not 0, pBindInfo must be a valid pointer to an array of bindInfoCount valid VkBindSparseInfo structures

If fence is not VK_NULL_HANDLE, fence must be a valid VkFence handle

The queue must support sparse binding operations

Both of fence, and queue that are valid handles must have been created, allocated, or retrieved from the same VkDevice

Host Synchronization

Host access to queue must be externally synchronized

Host access to pBindInfo[].pWaitSemaphores[] must be externally synchronized

Host access to pBindInfo[].pSignalSemaphores[] must be externally synchronized

Host access to pBindInfo[].pBufferBinds[].buffer must be externally synchronized

Host access to pBindInfo[].pImageOpaqueBinds[].image must be externally synchronized

Host access to pBindInfo[].pImageBinds[].image must be externally synchronized

Host access to fence must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| -               | -               | SPARSE_BINDING  | -               |
+-----------------+-----------------+-----------------+-----------------+

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_DEVICE_LOST

See Also

VkBindSparseInfo, VkFence, VkQueue