Graphics.Vulkan.C.Core11.Promoted_From_VK_KHR_protected_memory

Documentation

data VkDeviceQueueInfo2

VkDeviceQueueInfo2 - Structure specifying the parameters used for device queue creation

Description

The queue returned by vkGetDeviceQueue2 must have the same flags value from this structure as that used at device creation time in a VkDeviceQueueCreateInfo instance. If no matching flags were specified at device creation time then pQueue will return VK_NULL_HANDLE.

Valid Usage (Implicit)

See Also

VkDeviceQueueCreateFlags, VkStructureType, vkGetDeviceQueue2

Constructors

VkDeviceQueueInfo2

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DEVICE_QUEUE_INFO_2

vkPNext :: Ptr ()

pNext must be NULL

vkFlags :: VkDeviceQueueCreateFlags

flags must not be 0

vkQueueFamilyIndex :: Word32

queueFamilyIndex must be one of the queue family indices specified when device was created, via the VkDeviceQueueCreateInfo structure

vkQueueIndex :: Word32

queueIndex must be less than the number of queues created for the specified queue family index and VkDeviceQueueCreateFlags member flags equal to this flags value when device was created, via the queueCount member of the VkDeviceQueueCreateInfo structure

data VkPhysicalDeviceProtectedMemoryFeatures

VkPhysicalDeviceProtectedMemoryFeatures - Structure describing protected memory features that can be supported by an implementation

Description

If the VkPhysicalDeviceProtectedMemoryFeatures structure is included in the pNext chain of VkPhysicalDeviceFeatures2, it is filled with a value indicating whether the feature is supported.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

VkPhysicalDeviceProtectedMemoryFeatures

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PROTECTED_MEMORY_FEATURES

vkPNext :: Ptr ()

vkProtectedMemory :: VkBool32

protectedMemory specifies whether protected memory is supported.

data VkPhysicalDeviceProtectedMemoryProperties

VkPhysicalDeviceProtectedMemoryProperties - Structure describing protected memory properties that can be supported by an implementation

Description

If the VkPhysicalDeviceProtectedMemoryProperties structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with a value indicating the implementation-dependent behavior.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

VkPhysicalDeviceProtectedMemoryProperties

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PROTECTED_MEMORY_PROPERTIES

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkProtectedNoFault :: VkBool32

protectedNoFault specifies the behavior of the implementation when protected memory access rules are broken. If protectedNoFault is VK_TRUE, breaking those rules will not result in process termination or device loss.

data VkProtectedSubmitInfo

VkProtectedSubmitInfo - Structure indicating whether the submission is protected

Valid Usage

If the protected memory feature is not enabled, protectedSubmit must not be VK_TRUE.

If protectedSubmit is VK_TRUE, then each element of the pCommandBuffers array must be a protected command buffer.

If protectedSubmit is VK_FALSE, then each element of the pCommandBuffers array must be an unprotected command buffer.

If the VkSubmitInfo::pNext chain does not include a VkProtectedSubmitInfo structure, then each element of the command buffer of the pCommandBuffers array must be an unprotected command buffer.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PROTECTED_SUBMIT_INFO

See Also

VkBool32, VkStructureType

Constructors

VkProtectedSubmitInfo

Fields

vkSType :: VkStructureType

vkPNext :: Ptr ()

vkProtectedSubmit :: VkBool32

protectedSubmit specifies whether the batch is protected. If protectedSubmit is VK_TRUE, the batch is protected. If protectedSubmit is VK_FALSE, the batch is unprotected. If the VkSubmitInfo::pNext chain does not contain this structure, the batch is unprotected.

type FN_vkGetDeviceQueue2 = ("device" ::: VkDevice) -> ("pQueueInfo" ::: Ptr VkDeviceQueueInfo2) -> ("pQueue" ::: Ptr VkQueue) -> IO ()

type PFN_vkGetDeviceQueue2 = FunPtr FN_vkGetDeviceQueue2

vkGetDeviceQueue2 :: DeviceCmds -> ("device" ::: VkDevice) -> ("pQueueInfo" ::: Ptr VkDeviceQueueInfo2) -> ("pQueue" ::: Ptr VkQueue) -> IO ()

vkGetDeviceQueue2 - Get a queue handle from a device

Parameters

device is the logical device that owns the queue.

pQueueInfo points to an instance of the VkDeviceQueueInfo2 structure, describing the parameters used to create the device queue.

pQueue is a pointer to a VkQueue object that will be filled with the handle for the requested queue.

Valid Usage (Implicit)

See Also

VkDevice, VkDeviceQueueInfo2, VkQueue

pattern VK_BUFFER_CREATE_PROTECTED_BIT :: VkBufferCreateFlagBits

pattern VK_COMMAND_POOL_CREATE_PROTECTED_BIT :: VkCommandPoolCreateFlagBits

pattern VK_DEVICE_QUEUE_CREATE_PROTECTED_BIT :: VkDeviceQueueCreateFlagBits

VK_DEVICE_QUEUE_CREATE_PROTECTED_BIT specifies that the device queue is a protected-capable queue. If the protected memory feature is not enabled, the VK_DEVICE_QUEUE_CREATE_PROTECTED_BIT bit of flags must not be set.

pattern VK_IMAGE_CREATE_PROTECTED_BIT :: VkImageCreateFlagBits

pattern VK_MEMORY_PROPERTY_PROTECTED_BIT :: VkMemoryPropertyFlagBits

pattern VK_QUEUE_PROTECTED_BIT :: VkQueueFlagBits

pattern VK_STRUCTURE_TYPE_DEVICE_QUEUE_INFO_2 :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PROTECTED_MEMORY_FEATURES :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PROTECTED_MEMORY_PROPERTIES :: VkStructureType

pattern VK_STRUCTURE_TYPE_PROTECTED_SUBMIT_INFO :: VkStructureType