Graphics.Vulkan.C.Core11.Promoted_From_VK_KHR_subgroup

Documentation

data VkPhysicalDeviceSubgroupProperties

VkPhysicalDeviceSubgroupProperties - Structure describing subgroup support for an implementation

Members

The members of the VkPhysicalDeviceSubgroupProperties structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceSubgroupProperties structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkBool32, VkShaderStageFlags, VkStructureType, VkSubgroupFeatureFlags

Constructors

VkPhysicalDeviceSubgroupProperties

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SUBGROUP_PROPERTIES

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkSubgroupSize :: Word32

subgroupSize is the number of invocations in each subgroup. This will match any SubgroupSize decorated variable used in any shader module created on this device. subgroupSize is at least 1 if any of the physical device’s queues support VK_QUEUE_GRAPHICS_BIT or VK_QUEUE_COMPUTE_BIT.

vkSupportedStages :: VkShaderStageFlags

supportedStages is a bitfield of VkShaderStageFlagBits describing the shader stages that subgroup operations are supported in. supportedStages will have the VK_SHADER_STAGE_COMPUTE_BIT bit set if any of the physical device’s queues support VK_QUEUE_COMPUTE_BIT.

vkSupportedOperations :: VkSubgroupFeatureFlags

supportedOperations is a bitmask of VkSubgroupFeatureFlagBits specifying the sets of subgroup operations supported on this device. supportedOperations will have the VK_SUBGROUP_FEATURE_BASIC_BIT bit set if any of the physical device’s queues support VK_QUEUE_GRAPHICS_BIT or VK_QUEUE_COMPUTE_BIT.

vkQuadOperationsInAllStages :: VkBool32

quadOperationsInAllStages is a boolean that specifies whether quad subgroup operations are available in all stages, or are restricted to fragment and compute stages.

newtype VkSubgroupFeatureFlagBits

VkSubgroupFeatureFlagBits - Enum describing what subgroup operations are supported

See Also

VkSubgroupFeatureFlags

Constructors

VkSubgroupFeatureFlagBits VkFlags

pattern VK_SUBGROUP_FEATURE_BASIC_BIT :: VkSubgroupFeatureFlagBits

VK_SUBGROUP_FEATURE_BASIC_BIT specifies the device will accept SPIR-V shader modules that contain the GroupNonUniform capability.

pattern VK_SUBGROUP_FEATURE_VOTE_BIT :: VkSubgroupFeatureFlagBits

VK_SUBGROUP_FEATURE_VOTE_BIT specifies the device will accept SPIR-V shader modules that contain the GroupNonUniformVote capability.

pattern VK_SUBGROUP_FEATURE_ARITHMETIC_BIT :: VkSubgroupFeatureFlagBits

VK_SUBGROUP_FEATURE_ARITHMETIC_BIT specifies the device will accept SPIR-V shader modules that contain the GroupNonUniformArithmetic capability.

pattern VK_SUBGROUP_FEATURE_BALLOT_BIT :: VkSubgroupFeatureFlagBits

VK_SUBGROUP_FEATURE_BALLOT_BIT specifies the device will accept SPIR-V shader modules that contain the GroupNonUniformBallot capability.

pattern VK_SUBGROUP_FEATURE_SHUFFLE_BIT :: VkSubgroupFeatureFlagBits

VK_SUBGROUP_FEATURE_SHUFFLE_BIT specifies the device will accept SPIR-V shader modules that contain the GroupNonUniformShuffle capability.

pattern VK_SUBGROUP_FEATURE_SHUFFLE_RELATIVE_BIT :: VkSubgroupFeatureFlagBits

VK_SUBGROUP_FEATURE_SHUFFLE_RELATIVE_BIT specifies the device will accept SPIR-V shader modules that contain the GroupNonUniformShuffleRelative capability.

pattern VK_SUBGROUP_FEATURE_CLUSTERED_BIT :: VkSubgroupFeatureFlagBits

VK_SUBGROUP_FEATURE_CLUSTERED_BIT specifies the device will accept SPIR-V shader modules that contain the GroupNonUniformClustered capability.

pattern VK_SUBGROUP_FEATURE_QUAD_BIT :: VkSubgroupFeatureFlagBits

VK_SUBGROUP_FEATURE_QUAD_BIT specifies the device will accept SPIR-V shader modules that contain the GroupNonUniformQuad capability.

type VkSubgroupFeatureFlags = VkSubgroupFeatureFlagBits

VkSubgroupFeatureFlags - Bitmask of VkSubgroupFeatureFlagBits

Description

VkSubgroupFeatureFlags is a bitmask type for setting a mask of zero or more VkSubgroupFeatureFlagBits.

See Also

VkPhysicalDeviceSubgroupProperties, VkSubgroupFeatureFlagBits

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SUBGROUP_PROPERTIES :: VkStructureType