Graphics.Vulkan.C.Core11.Promoted_from_VK_KHR_16bit_storage

Documentation

data VkPhysicalDevice16BitStorageFeatures

VkPhysicalDevice16BitStorageFeatures - Structure describing features supported by VK_KHR_16bit_storage

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

VkPhysicalDevice16BitStorageFeatures

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_16BIT_STORAGE_FEATURES

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkStorageBuffer16BitAccess :: VkBool32

storageBuffer16BitAccess specifies whether objects in the StorageBuffer storage class with the Block decoration can have 16-bit integer and 16-bit floating-point members. If this feature is not enabled, 16-bit integer or 16-bit floating-point members must not be used in such objects. This also specifies whether shader modules can declare the StorageBuffer16BitAccess capability.

vkUniformAndStorageBuffer16BitAccess :: VkBool32

uniformAndStorageBuffer16BitAccess specifies whether objects in the Uniform storage class with the Block decoration and in the StorageBuffer storage class with the same decoration can have 16-bit integer and 16-bit floating-point members. If this feature is not enabled, 16-bit integer or 16-bit floating-point members must not be used in such objects. This also specifies whether shader modules can declare the UniformAndStorageBuffer16BitAccess capability.

vkStoragePushConstant16 :: VkBool32

storagePushConstant16 specifies whether objects in the PushConstant storage class can have 16-bit integer and 16-bit floating-point members. If this feature is not enabled, 16-bit integer or floating-point members must not be used in such objects. This also specifies whether shader modules can declare the StoragePushConstant16 capability.

vkStorageInputOutput16 :: VkBool32

storageInputOutput16 specifies whether objects in the Input and Output storage classes can have 16-bit integer and 16-bit floating-point members. If this feature is not enabled, 16-bit integer or 16-bit floating-point members must not be used in such objects. This also specifies whether shader modules can declare the StorageInputOutput16 capability.

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_16BIT_STORAGE_FEATURES :: VkStructureType