Graphics.Vulkan.C.Core11.Promoted_from_VK_KHR_bind_memory2

Documentation

data VkBindBufferMemoryInfo

VkBindBufferMemoryInfo - Structure specifying how to bind a buffer to memory

Valid Usage

buffer must not already be backed by a memory object

buffer must not have been created with any sparse memory binding flags

memoryOffset must be less than the size of memory

memory must have been allocated using one of the memory types allowed in the memoryTypeBits member of the VkMemoryRequirements structure returned from a call to vkGetBufferMemoryRequirements with buffer

memoryOffset must be an integer multiple of the alignment member of the VkMemoryRequirements structure returned from a call to vkGetBufferMemoryRequirements with buffer

The size member of the VkMemoryRequirements structure returned from a call to vkGetBufferMemoryRequirements with buffer must be less than or equal to the size of memory minus memoryOffset

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_BIND_BUFFER_MEMORY_INFO

pNext must be NULL or a pointer to a valid instance of VkBindBufferMemoryDeviceGroupInfo

buffer must be a valid VkBuffer handle

memory must be a valid VkDeviceMemory handle

Both of buffer, and memory must have been created, allocated, or retrieved from the same VkDevice

See Also

VkBuffer, VkDeviceMemory, VkDeviceSize, VkStructureType, vkBindBufferMemory2, vkBindBufferMemory2KHR

Constructors

VkBindBufferMemoryInfo

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkBuffer :: VkBuffer

buffer is the buffer to be attached to memory.

vkMemory :: VkDeviceMemory

memory is a VkDeviceMemory object describing the device memory to attach.

vkMemoryOffset :: VkDeviceSize

memoryOffset is the start offset of the region of memory which is to be bound to the buffer. The number of bytes returned in the VkMemoryRequirements::size member in memory, starting from memoryOffset bytes, will be bound to the specified buffer.

data VkBindImageMemoryInfo

VkBindImageMemoryInfo - Structure specifying how to bind an image to memory

Valid Usage

image must not already be backed by a memory object

image must not have been created with any sparse memory binding flags

memoryOffset must be less than the size of memory

memory must have been allocated using one of the memory types allowed in the memoryTypeBits member of the VkMemoryRequirements structure returned from a call to vkGetImageMemoryRequirements with image

memoryOffset must be an integer multiple of the alignment member of the VkMemoryRequirements structure returned from a call to vkGetImageMemoryRequirements with image

The difference of the size of memory and memoryOffset must be greater than or equal to the size member of the VkMemoryRequirements structure returned from a call to vkGetImageMemoryRequirements with the same image

memory must be a valid VkDeviceMemory handle

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_BIND_IMAGE_MEMORY_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkBindImageMemoryDeviceGroupInfo, VkBindImageMemorySwapchainInfoKHR, or VkBindImagePlaneMemoryInfo

Each sType member in the pNext chain must be unique

image must be a valid VkImage handle

Both of image, and memory that are valid handles must have been created, allocated, or retrieved from the same VkDevice

See Also

VkDeviceMemory, VkDeviceSize, VkImage, VkStructureType, vkBindImageMemory2, vkBindImageMemory2KHR

Constructors

VkBindImageMemoryInfo

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkImage :: VkImage

image is the image to be attached to memory.

vkMemory :: VkDeviceMemory

memory is a VkDeviceMemory object describing the device memory to attach.

vkMemoryOffset :: VkDeviceSize

memoryOffset is the start offset of the region of memory which is to be bound to the image. The number of bytes returned in the VkMemoryRequirements::size member in memory, starting from memoryOffset bytes, will be bound to the specified image.

type FN_vkBindBufferMemory2 = ("device" ::: VkDevice) -> ("bindInfoCount" ::: Word32) -> ("pBindInfos" ::: Ptr VkBindBufferMemoryInfo) -> IO VkResult

type PFN_vkBindBufferMemory2 = FunPtr FN_vkBindBufferMemory2

vkBindBufferMemory2 :: DeviceCmds -> ("device" ::: VkDevice) -> ("bindInfoCount" ::: Word32) -> ("pBindInfos" ::: Ptr VkBindBufferMemoryInfo) -> IO VkResult

vkBindBufferMemory2 - Bind device memory to buffer objects

Parameters

device is the logical device that owns the buffers and memory.

bindInfoCount is the number of elements in pBindInfos.

pBindInfos is a pointer to an array of structures of type VkBindBufferMemoryInfo, describing buffers and memory to bind.

Description

On some implementations, it may be more efficient to batch memory bindings into a single command.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkBindBufferMemoryInfo, VkDevice

type FN_vkBindImageMemory2 = ("device" ::: VkDevice) -> ("bindInfoCount" ::: Word32) -> ("pBindInfos" ::: Ptr VkBindImageMemoryInfo) -> IO VkResult

type PFN_vkBindImageMemory2 = FunPtr FN_vkBindImageMemory2

vkBindImageMemory2 :: DeviceCmds -> ("device" ::: VkDevice) -> ("bindInfoCount" ::: Word32) -> ("pBindInfos" ::: Ptr VkBindImageMemoryInfo) -> IO VkResult

vkBindImageMemory2 - Bind device memory to image objects

Parameters

device is the logical device that owns the images and memory.

bindInfoCount is the number of elements in pBindInfos.

pBindInfos is a pointer to an array of structures of type VkBindImageMemoryInfo, describing images and memory to bind.

Description

On some implementations, it may be more efficient to batch memory bindings into a single command.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkBindImageMemoryInfo, VkDevice

pattern VK_IMAGE_CREATE_ALIAS_BIT :: VkImageCreateFlagBits

pattern VK_STRUCTURE_TYPE_BIND_BUFFER_MEMORY_INFO :: VkStructureType

pattern VK_STRUCTURE_TYPE_BIND_IMAGE_MEMORY_INFO :: VkStructureType