Graphics.Vulkan.C.Core11.Promoted_from_VK_KHR_external_fence

Documentation

data VkExportFenceCreateInfo

VkExportFenceCreateInfo - Structure specifying handle types that can be exported from a fence

Valid Usage

The bits in handleTypes must be supported and compatible, as reported by VkExternalFenceProperties.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_EXPORT_FENCE_CREATE_INFO

handleTypes must be a valid combination of VkExternalFenceHandleTypeFlagBits values

See Also

VkExternalFenceHandleTypeFlags, VkStructureType

Constructors

VkExportFenceCreateInfo

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkHandleTypes :: VkExternalFenceHandleTypeFlags

handleTypes is a bitmask of VkExternalFenceHandleTypeFlagBits specifying one or more fence handle types the application can export from the resulting fence. The application can request multiple handle types for the same fence.

newtype VkFenceImportFlagBits

VkFenceImportFlagBits - Bitmask specifying additional parameters of fence payload import

See Also

VkFenceImportFlags

Constructors

VkFenceImportFlagBits VkFlags

pattern VK_FENCE_IMPORT_TEMPORARY_BIT :: VkFenceImportFlagBits

VK_FENCE_IMPORT_TEMPORARY_BIT specifies that the fence payload will be imported only temporarily, as described in Importing Fence Payloads, regardless of the permanence of handleType.

type VkFenceImportFlags = VkFenceImportFlagBits

VkFenceImportFlags - Bitmask of VkFenceImportFlagBits

Description

VkFenceImportFlags is a bitmask type for setting a mask of zero or more VkFenceImportFlagBits.

See Also

VkFenceImportFlagBits, VkImportFenceFdInfoKHR, VkImportFenceWin32HandleInfoKHR

pattern VK_STRUCTURE_TYPE_EXPORT_FENCE_CREATE_INFO :: VkStructureType