Graphics.Vulkan.C.Core11.Promoted_from_VK_KHR_external_memory_capabilities

Documentation

data VkExternalBufferProperties

VkExternalBufferProperties - Structure specifying supported external handle capabilities

Valid Usage (Implicit)

See Also

VkExternalMemoryProperties, VkStructureType, vkGetPhysicalDeviceExternalBufferProperties, vkGetPhysicalDeviceExternalBufferPropertiesKHR

Constructors

VkExternalBufferProperties

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_EXTERNAL_BUFFER_PROPERTIES

vkPNext :: Ptr ()

pNext must be NULL

vkExternalMemoryProperties :: VkExternalMemoryProperties

externalMemoryProperties is an instance of the VkExternalMemoryProperties structure specifying various capabilities of the external handle type when used with the specified buffer creation parameters.

data VkExternalImageFormatProperties

VkExternalImageFormatProperties - Structure specifying supported external handle properties

Valid Usage (Implicit)

See Also

VkExternalMemoryProperties, VkStructureType

Constructors

VkExternalImageFormatProperties

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_EXTERNAL_IMAGE_FORMAT_PROPERTIES

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkExternalMemoryProperties :: VkExternalMemoryProperties

externalMemoryProperties is an instance of the VkExternalMemoryProperties structure specifying various capabilities of the external handle type when used with the specified image creation parameters.

newtype VkExternalMemoryFeatureFlagBits

VkExternalMemoryFeatureFlagBits - Bitmask specifying features of an external memory handle type

Description

Because their semantics in external APIs roughly align with that of an image or buffer with a dedicated allocation in Vulkan, implementations are required to report VK_EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT for the following external handle types:

See Also

VkExternalMemoryFeatureFlags

Constructors

VkExternalMemoryFeatureFlagBits VkFlags

pattern VK_EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT :: VkExternalMemoryFeatureFlagBits

VK_EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT specifies that images or buffers created with the specified parameters and handle type must use the mechanisms defined by VkMemoryDedicatedRequirements and VkMemoryDedicatedAllocateInfo to create (or import) a dedicated allocation for the image or buffer.

pattern VK_EXTERNAL_MEMORY_FEATURE_EXPORTABLE_BIT :: VkExternalMemoryFeatureFlagBits

VK_EXTERNAL_MEMORY_FEATURE_EXPORTABLE_BIT specifies that handles of this type can be exported from Vulkan memory objects.

pattern VK_EXTERNAL_MEMORY_FEATURE_IMPORTABLE_BIT :: VkExternalMemoryFeatureFlagBits

VK_EXTERNAL_MEMORY_FEATURE_IMPORTABLE_BIT specifies that handles of this type can be imported as Vulkan memory objects.

type VkExternalMemoryFeatureFlags = VkExternalMemoryFeatureFlagBits

VkExternalMemoryFeatureFlags - Bitmask of VkExternalMemoryFeatureFlagBits

Description

VkExternalMemoryFeatureFlags is a bitmask type for setting a mask of zero or more VkExternalMemoryFeatureFlagBits.

See Also

VkExternalMemoryFeatureFlagBits, VkExternalMemoryProperties

newtype VkExternalMemoryHandleTypeFlagBits

VkExternalMemoryHandleTypeFlagBits - Bit specifying external memory handle types

Description

Some external memory handle types can only be shared within the same underlying physical device and/or the same driver version, as defined in the following table:

+----------------------+----------------------+-----------------------+
| Handle type          | 'VkPhysicalDeviceIDP | 'VkPhysicalDeviceIDPr |
|                      | roperties'::@driverU | operties'::@deviceUUI |
|                      | UID@                 | D@                    |
+----------------------+----------------------+-----------------------+
| 'VK_EXTERNAL_MEMORY_ | Must match           | Must match            |
| HANDLE_TYPE_OPAQUE_F |                      |                       |
| D_BIT'               |                      |                       |
+----------------------+----------------------+-----------------------+
| 'VK_EXTERNAL_MEMORY_ | Must match           | Must match            |
| HANDLE_TYPE_OPAQUE_W |                      |                       |
| IN32_BIT'            |                      |                       |
+----------------------+----------------------+-----------------------+
| 'VK_EXTERNAL_MEMORY_ | Must match           | Must match            |
| HANDLE_TYPE_OPAQUE_W |                      |                       |
| IN32_KMT_BIT'        |                      |                       |
+----------------------+----------------------+-----------------------+
| 'VK_EXTERNAL_MEMORY_ | Must match           | Must match            |
| HANDLE_TYPE_D3D11_TE |                      |                       |
| XTURE_BIT'           |                      |                       |
+----------------------+----------------------+-----------------------+
| 'VK_EXTERNAL_MEMORY_ | Must match           | Must match            |
| HANDLE_TYPE_D3D11_TE |                      |                       |
| XTURE_KMT_BIT'       |                      |                       |
+----------------------+----------------------+-----------------------+
| 'VK_EXTERNAL_MEMORY_ | Must match           | Must match            |
| HANDLE_TYPE_D3D12_HE |                      |                       |
| AP_BIT'              |                      |                       |
+----------------------+----------------------+-----------------------+
| 'VK_EXTERNAL_MEMORY_ | Must match           | Must match            |
| HANDLE_TYPE_D3D12_RE |                      |                       |
| SOURCE_BIT'          |                      |                       |
+----------------------+----------------------+-----------------------+

External memory handle types compatibility

See Also

VkExternalMemoryHandleTypeFlags, VkImportMemoryFdInfoKHR, VkImportMemoryHostPointerInfoEXT, VkImportMemoryWin32HandleInfoKHR, VkMemoryGetFdInfoKHR, VkMemoryGetWin32HandleInfoKHR, VkPhysicalDeviceExternalBufferInfo, VkPhysicalDeviceExternalImageFormatInfo, vkGetMemoryFdPropertiesKHR, vkGetMemoryHostPointerPropertiesEXT, vkGetMemoryWin32HandlePropertiesKHR

Constructors

VkExternalMemoryHandleTypeFlagBits VkFlags

pattern VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_FD_BIT :: VkExternalMemoryHandleTypeFlagBits

VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_FD_BIT specifies a POSIX file descriptor handle that has only limited valid usage outside of Vulkan and other compatible APIs. It must be compatible with the POSIX system calls dup, dup2, close, and the non-standard system call dup3. Additionally, it must be transportable over a socket using an SCM_RIGHTS control message. It owns a reference to the underlying memory resource represented by its Vulkan memory object.

pattern VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT :: VkExternalMemoryHandleTypeFlagBits

VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT specifies an NT handle that has only limited valid usage outside of Vulkan and other compatible APIs. It must be compatible with the functions DuplicateHandle, CloseHandle, CompareObjectHandles, GetHandleInformation, and SetHandleInformation. It owns a reference to the underlying memory resource represented by its Vulkan memory object.

pattern VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT :: VkExternalMemoryHandleTypeFlagBits

VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT specifies a global share handle that has only limited valid usage outside of Vulkan and other compatible APIs. It is not compatible with any native APIs. It does not own a reference to the underlying memory resource represented its Vulkan memory object, and will therefore become invalid when all Vulkan memory objects associated with it are destroyed.

pattern VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_BIT :: VkExternalMemoryHandleTypeFlagBits

VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_BIT specifies an NT handle returned by IDXGIResource1::CreateSharedHandle referring to a Direct3D 10 or 11 texture resource. It owns a reference to the memory used by the Direct3D resource.

pattern VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_KMT_BIT :: VkExternalMemoryHandleTypeFlagBits

VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_KMT_BIT specifies a global share handle returned by IDXGIResource::GetSharedHandle referring to a Direct3D 10 or 11 texture resource. It does not own a reference to the underlying Direct3D resource, and will therefore become invalid when all Vulkan memory objects and Direct3D resources associated with it are destroyed.

pattern VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_HEAP_BIT :: VkExternalMemoryHandleTypeFlagBits

VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_HEAP_BIT specifies an NT handle returned by ID3D12Device::CreateSharedHandle referring to a Direct3D 12 heap resource. It owns a reference to the resources used by the Direct3D heap.

pattern VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_RESOURCE_BIT :: VkExternalMemoryHandleTypeFlagBits

VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_RESOURCE_BIT specifies an NT handle returned by ID3D12Device::CreateSharedHandle referring to a Direct3D 12 committed resource. It owns a reference to the memory used by the Direct3D resource.

type VkExternalMemoryHandleTypeFlags = VkExternalMemoryHandleTypeFlagBits

VkExternalMemoryHandleTypeFlags - Bitmask of VkExternalMemoryHandleTypeFlagBits

Description

VkExternalMemoryHandleTypeFlags is a bitmask type for setting a mask of zero or more VkExternalMemoryHandleTypeFlagBits.

See Also

VkExportMemoryAllocateInfo, VkExternalMemoryBufferCreateInfo, VkExternalMemoryHandleTypeFlagBits, VkExternalMemoryImageCreateInfo, VkExternalMemoryProperties

data VkExternalMemoryProperties

VkExternalMemoryProperties - Structure specifying external memory handle type capabilities

Description

compatibleHandleTypes must include at least handleType. Inclusion of a handle type in compatibleHandleTypes does not imply the values returned in VkImageFormatProperties2 will be the same when VkPhysicalDeviceExternalImageFormatInfo::handleType is set to that type. The application is responsible for querying the capabilities of all handle types intended for concurrent use in a single image and intersecting them to obtain the compatible set of capabilities.

See Also

VkExternalBufferProperties, VkExternalImageFormatProperties, VkExternalMemoryFeatureFlags, VkExternalMemoryHandleTypeFlags

Constructors

VkExternalMemoryProperties

Fields

vkExternalMemoryFeatures :: VkExternalMemoryFeatureFlags

externalMemoryFeatures is a bitmask of VkExternalMemoryFeatureFlagBits specifying the features of handleType.

vkExportFromImportedHandleTypes :: VkExternalMemoryHandleTypeFlags

exportFromImportedHandleTypes is a bitmask of VkExternalMemoryHandleTypeFlagBits specifying which types of imported handle handleType can be exported from.

vkCompatibleHandleTypes :: VkExternalMemoryHandleTypeFlags

compatibleHandleTypes is a bitmask of VkExternalMemoryHandleTypeFlagBits specifying handle types which can be specified at the same time as handleType when creating an image compatible with external memory.

data VkPhysicalDeviceExternalBufferInfo

VkPhysicalDeviceExternalBufferInfo - Structure specifying buffer creation parameters

Valid Usage (Implicit)

See Also

VkBufferCreateFlags, VkBufferUsageFlags, VkExternalMemoryHandleTypeFlagBits, VkStructureType, vkGetPhysicalDeviceExternalBufferProperties, vkGetPhysicalDeviceExternalBufferPropertiesKHR

Constructors

VkPhysicalDeviceExternalBufferInfo

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_BUFFER_INFO

vkPNext :: Ptr ()

pNext must be NULL

vkFlags :: VkBufferCreateFlags

flags must be a valid combination of VkBufferCreateFlagBits values

vkUsage :: VkBufferUsageFlags

usage must not be 0

vkHandleType :: VkExternalMemoryHandleTypeFlagBits

handleType must be a valid VkExternalMemoryHandleTypeFlagBits value

data VkPhysicalDeviceExternalImageFormatInfo

VkPhysicalDeviceExternalImageFormatInfo - Structure specifying external image creation parameters

Description

If handleType is 0, vkGetPhysicalDeviceImageFormatProperties2 will behave as if VkPhysicalDeviceExternalImageFormatInfo was not present, and VkExternalImageFormatProperties will be ignored.

If handleType is not compatible with the format, type, tiling, usage, and flags specified in VkPhysicalDeviceImageFormatInfo2, then vkGetPhysicalDeviceImageFormatProperties2 returns VK_ERROR_FORMAT_NOT_SUPPORTED.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_IMAGE_FORMAT_INFO

If handleType is not 0, handleType must be a valid VkExternalMemoryHandleTypeFlagBits value

See Also

VkExternalMemoryHandleTypeFlagBits, VkStructureType

Constructors

VkPhysicalDeviceExternalImageFormatInfo

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkHandleType :: VkExternalMemoryHandleTypeFlagBits

handleType is a VkExternalMemoryHandleTypeFlagBits value specifying the memory handle type that will be used with the memory associated with the image.

data VkPhysicalDeviceIDProperties

VkPhysicalDeviceIDProperties - Structure specifying IDs related to the physical device

Description

deviceUUID must be immutable for a given device across instances, processes, driver APIs, driver versions, and system reboots.

Applications can compare the driverUUID value across instance and process boundaries, and can make similar queries in external APIs to determine whether they are capable of sharing memory objects and resources using them with the device.

deviceUUID and/or driverUUID must be used to determine whether a particular external object can be shared between driver components, where such a restriction exists as defined in the compatibility table for the particular object type:

If deviceLUIDValid is VK_FALSE, the values of deviceLUID and deviceNodeMask are undefined. If deviceLUIDValid is VK_TRUE and Vulkan is running on the Windows operating system, the contents of deviceLUID can be cast to an LUID object and must be equal to the locally unique identifier of a IDXGIAdapter1 object that corresponds to physicalDevice. If deviceLUIDValid is VK_TRUE, deviceNodeMask must contain exactly one bit. If Vulkan is running on an operating system that supports the Direct3D 12 API and physicalDevice corresponds to an individual device in a linked device adapter, deviceNodeMask identifies the Direct3D 12 node corresponding to physicalDevice. Otherwise, deviceNodeMask must be 1.

Note

Although they have identical descriptions, VkPhysicalDeviceIDProperties::deviceUUID may differ from VkPhysicalDeviceProperties2::pipelineCacheUUID. The former is intended to identify and correlate devices across API and driver boundaries, while the latter is used to identify a compatible device and driver combination to use when serializing and de-serializing pipeline state.

Note

While VkPhysicalDeviceIDProperties::deviceUUID is specified to remain consistent across driver versions and system reboots, it is not intended to be usable as a serializable persistent identifier for a device. It may change when a device is physically added to, removed from, or moved to a different connector in a system while that system is powered down. Further, there is no reasonable way to verify with conformance testing that a given device retains the same UUID in a given system across all driver versions supported in that system. While implementations should make every effort to report consistent device UUIDs across driver versions, applications should avoid relying on the persistence of this value for uses other than identifying compatible devices for external object sharing purposes.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

VkPhysicalDeviceIDProperties

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_ID_PROPERTIES

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkDeviceUUID :: Vector VK_UUID_SIZE Word8

deviceUUID is an array of size VK_UUID_SIZE, containing 8-bit values that represent a universally unique identifier for the device.

vkDriverUUID :: Vector VK_UUID_SIZE Word8

driverUUID is an array of size VK_UUID_SIZE, containing 8-bit values that represent a universally unique identifier for the driver build in use by the device.

vkDeviceLUID :: Vector VK_LUID_SIZE Word8

deviceLUID is an array of size VK_LUID_SIZE, containing 8-bit values that represent a locally unique identifier for the device.

vkDeviceNodeMask :: Word32

deviceNodeMask is a bitfield identifying the node within a linked device adapter corresponding to the device.

vkDeviceLUIDValid :: VkBool32

deviceLUIDValid is a boolean value that will be VK_TRUE if deviceLUID contains a valid LUID and deviceNodeMask contains a valid node mask, and VK_FALSE if they do not.

type FN_vkGetPhysicalDeviceExternalBufferProperties = ("physicalDevice" ::: VkPhysicalDevice) -> ("pExternalBufferInfo" ::: Ptr VkPhysicalDeviceExternalBufferInfo) -> ("pExternalBufferProperties" ::: Ptr VkExternalBufferProperties) -> IO ()

type PFN_vkGetPhysicalDeviceExternalBufferProperties = FunPtr FN_vkGetPhysicalDeviceExternalBufferProperties

vkGetPhysicalDeviceExternalBufferProperties :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pExternalBufferInfo" ::: Ptr VkPhysicalDeviceExternalBufferInfo) -> ("pExternalBufferProperties" ::: Ptr VkExternalBufferProperties) -> IO ()

vkGetPhysicalDeviceExternalBufferProperties - Query external handle types supported by buffers

Parameters

physicalDevice is the physical device from which to query the buffer capabilities.

pExternalBufferInfo points to an instance of the VkPhysicalDeviceExternalBufferInfo structure, describing the parameters that would be consumed by vkCreateBuffer.

pExternalBufferProperties points to an instance of the VkExternalBufferProperties structure in which capabilities are returned.

Valid Usage (Implicit)

See Also

VkExternalBufferProperties, VkPhysicalDevice, VkPhysicalDeviceExternalBufferInfo

type VK_LUID_SIZE = 8

pattern VK_LUID_SIZE :: Integral a => a

pattern VK_STRUCTURE_TYPE_EXTERNAL_BUFFER_PROPERTIES :: VkStructureType

pattern VK_STRUCTURE_TYPE_EXTERNAL_IMAGE_FORMAT_PROPERTIES :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_BUFFER_INFO :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_IMAGE_FORMAT_INFO :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_ID_PROPERTIES :: VkStructureType