Graphics.Vulkan.C.Core11.Promoted_from_VK_KHR_external_semaphore_capabilities

Documentation

newtype VkExternalSemaphoreFeatureFlagBits

VkExternalSemaphoreFeatureFlagBits - Bitfield describing features of an external semaphore handle type

See Also

VkExternalSemaphoreFeatureFlags

Constructors

VkExternalSemaphoreFeatureFlagBits VkFlags

pattern VK_EXTERNAL_SEMAPHORE_FEATURE_EXPORTABLE_BIT :: VkExternalSemaphoreFeatureFlagBits

VK_EXTERNAL_SEMAPHORE_FEATURE_EXPORTABLE_BIT specifies that handles of this type can be exported from Vulkan semaphore objects.

pattern VK_EXTERNAL_SEMAPHORE_FEATURE_IMPORTABLE_BIT :: VkExternalSemaphoreFeatureFlagBits

VK_EXTERNAL_SEMAPHORE_FEATURE_IMPORTABLE_BIT specifies that handles of this type can be imported as Vulkan semaphore objects.

type VkExternalSemaphoreFeatureFlags = VkExternalSemaphoreFeatureFlagBits

VkExternalSemaphoreFeatureFlags - Bitmask of VkExternalSemaphoreFeatureFlagBitsKHR

Description

VkExternalSemaphoreFeatureFlags is a bitmask type for setting a mask of zero or more VkExternalSemaphoreFeatureFlagBits.

See Also

VkExternalSemaphoreFeatureFlagBits, VkExternalSemaphoreProperties

newtype VkExternalSemaphoreHandleTypeFlagBits

VkExternalSemaphoreHandleTypeFlagBits - Bitmask of valid external semaphore handle types

Description

Note

Handles of type VK_EXTERNAL_SEMAPHORE_HANDLE_TYPE_SYNC_FD_BIT generated by the implementation may represent either Linux Sync Files or Android Fences at the implementation’s discretion. Applications should only use operations defined for both types of file descriptors, unless they know via means external to Vulkan the type of the file descriptor, or are prepared to deal with the system-defined operation failures resulting from using the wrong type.

Some external semaphore handle types can only be shared within the same underlying physical device and/or the same driver version, as defined in the following table:

+----------------------+----------------------+-----------------------+
| Handle type          | 'Graphics.Vulkan.C.C | 'Graphics.Vulkan.C.Co |
|                      | ore11.Promoted_from_ | re11.Promoted_from_VK |
|                      | VK_KHR_external_memo | _KHR_external_memory_ |
|                      | ry_capabilities.VkPh | capabilities.VkPhysic |
|                      | ysicalDeviceIDProper | alDeviceIDProperties' |
|                      | ties'::@driverUUID@  | ::@deviceUUID@        |
+----------------------+----------------------+-----------------------+
| 'VK_EXTERNAL_SEMAPHO | Must match           | Must match            |
| RE_HANDLE_TYPE_OPAQU |                      |                       |
| E_FD_BIT'            |                      |                       |
+----------------------+----------------------+-----------------------+
| 'VK_EXTERNAL_SEMAPHO | Must match           | Must match            |
| RE_HANDLE_TYPE_OPAQU |                      |                       |
| E_WIN32_BIT'         |                      |                       |
+----------------------+----------------------+-----------------------+
| 'VK_EXTERNAL_SEMAPHO | Must match           | Must match            |
| RE_HANDLE_TYPE_OPAQU |                      |                       |
| E_WIN32_KMT_BIT'     |                      |                       |
+----------------------+----------------------+-----------------------+
| 'VK_EXTERNAL_SEMAPHO | Must match           | Must match            |
| RE_HANDLE_TYPE_D3D12 |                      |                       |
| _FENCE_BIT'          |                      |                       |
+----------------------+----------------------+-----------------------+
| 'VK_EXTERNAL_SEMAPHO | No restriction       | No restriction        |
| RE_HANDLE_TYPE_SYNC_ |                      |                       |
| FD_BIT'              |                      |                       |
+----------------------+----------------------+-----------------------+

External semaphore handle types compatibility

See Also

VkExternalSemaphoreHandleTypeFlags, VkImportSemaphoreFdInfoKHR, VkImportSemaphoreWin32HandleInfoKHR, VkPhysicalDeviceExternalSemaphoreInfo, VkSemaphoreGetFdInfoKHR, VkSemaphoreGetWin32HandleInfoKHR

Constructors

VkExternalSemaphoreHandleTypeFlagBits VkFlags

pattern VK_EXTERNAL_SEMAPHORE_HANDLE_TYPE_OPAQUE_FD_BIT :: VkExternalSemaphoreHandleTypeFlagBits

VK_EXTERNAL_SEMAPHORE_HANDLE_TYPE_OPAQUE_FD_BIT specifies a POSIX file descriptor handle that has only limited valid usage outside of Vulkan and other compatible APIs. It must be compatible with the POSIX system calls dup, dup2, close, and the non-standard system call dup3. Additionally, it must be transportable over a socket using an SCM_RIGHTS control message. It owns a reference to the underlying synchronization primitive represented by its Vulkan semaphore object.

pattern VK_EXTERNAL_SEMAPHORE_HANDLE_TYPE_OPAQUE_WIN32_BIT :: VkExternalSemaphoreHandleTypeFlagBits

VK_EXTERNAL_SEMAPHORE_HANDLE_TYPE_OPAQUE_WIN32_BIT specifies an NT handle that has only limited valid usage outside of Vulkan and other compatible APIs. It must be compatible with the functions DuplicateHandle, CloseHandle, CompareObjectHandles, GetHandleInformation, and SetHandleInformation. It owns a reference to the underlying synchronization primitive represented by its Vulkan semaphore object.

pattern VK_EXTERNAL_SEMAPHORE_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT :: VkExternalSemaphoreHandleTypeFlagBits

VK_EXTERNAL_SEMAPHORE_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT specifies a global share handle that has only limited valid usage outside of Vulkan and other compatible APIs. It is not compatible with any native APIs. It does not own a reference to the underlying synchronization primitive represented its Vulkan semaphore object, and will therefore become invalid when all Vulkan semaphore objects associated with it are destroyed.

pattern VK_EXTERNAL_SEMAPHORE_HANDLE_TYPE_D3D12_FENCE_BIT :: VkExternalSemaphoreHandleTypeFlagBits

VK_EXTERNAL_SEMAPHORE_HANDLE_TYPE_D3D12_FENCE_BIT specifies an NT handle returned by ID3D12Device::CreateSharedHandle referring to a Direct3D 12 fence. It owns a reference to the underlying synchronization primitive associated with the Direct3D fence.

pattern VK_EXTERNAL_SEMAPHORE_HANDLE_TYPE_SYNC_FD_BIT :: VkExternalSemaphoreHandleTypeFlagBits

VK_EXTERNAL_SEMAPHORE_HANDLE_TYPE_SYNC_FD_BIT specifies a POSIX file descriptor handle to a Linux Sync File or Android Fence object. It can be used with any native API accepting a valid sync file or fence as input. It owns a reference to the underlying synchronization primitive associated with the file descriptor. Implementations which support importing this handle type must accept any type of sync or fence FD supported by the native system they are running on.

type VkExternalSemaphoreHandleTypeFlags = VkExternalSemaphoreHandleTypeFlagBits

VkExternalSemaphoreHandleTypeFlags - Bitmask of VkExternalSemaphoreHandleTypeFlagBits

Description

VkExternalSemaphoreHandleTypeFlags is a bitmask type for setting a mask of zero or more VkExternalSemaphoreHandleTypeFlagBits.

See Also

VkExportSemaphoreCreateInfo, VkExternalSemaphoreHandleTypeFlagBits, VkExternalSemaphoreProperties

data VkExternalSemaphoreProperties

VkExternalSemaphoreProperties - Structure describing supported external semaphore handle features

Description

If handleType is not supported by the implementation, then VkExternalSemaphoreProperties::externalSemaphoreFeatures will be set to zero.

Valid Usage (Implicit)

See Also

VkExternalSemaphoreFeatureFlags, VkExternalSemaphoreHandleTypeFlags, VkStructureType, vkGetPhysicalDeviceExternalSemaphoreProperties, vkGetPhysicalDeviceExternalSemaphorePropertiesKHR

Constructors

VkExternalSemaphoreProperties

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_EXTERNAL_SEMAPHORE_PROPERTIES

vkPNext :: Ptr ()

pNext must be NULL

vkExportFromImportedHandleTypes :: VkExternalSemaphoreHandleTypeFlags

exportFromImportedHandleTypes is a bitmask of VkExternalSemaphoreHandleTypeFlagBits specifying which types of imported handle handleType can be exported from.

vkCompatibleHandleTypes :: VkExternalSemaphoreHandleTypeFlags

compatibleHandleTypes is a bitmask of VkExternalSemaphoreHandleTypeFlagBits specifying handle types which can be specified at the same time as handleType when creating a semaphore.

vkExternalSemaphoreFeatures :: VkExternalSemaphoreFeatureFlags

externalSemaphoreFeatures is a bitmask of VkExternalSemaphoreFeatureFlagBits describing the features of handleType.

data VkPhysicalDeviceExternalSemaphoreInfo

VkPhysicalDeviceExternalSemaphoreInfo - Structure specifying semaphore creation parameters.

Valid Usage (Implicit)

See Also

VkExternalSemaphoreHandleTypeFlagBits, VkStructureType, vkGetPhysicalDeviceExternalSemaphoreProperties, vkGetPhysicalDeviceExternalSemaphorePropertiesKHR

Constructors

VkPhysicalDeviceExternalSemaphoreInfo

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_SEMAPHORE_INFO

vkPNext :: Ptr ()

pNext must be NULL

vkHandleType :: VkExternalSemaphoreHandleTypeFlagBits

handleType must be a valid VkExternalSemaphoreHandleTypeFlagBits value

type FN_vkGetPhysicalDeviceExternalSemaphoreProperties = ("physicalDevice" ::: VkPhysicalDevice) -> ("pExternalSemaphoreInfo" ::: Ptr VkPhysicalDeviceExternalSemaphoreInfo) -> ("pExternalSemaphoreProperties" ::: Ptr VkExternalSemaphoreProperties) -> IO ()

type PFN_vkGetPhysicalDeviceExternalSemaphoreProperties = FunPtr FN_vkGetPhysicalDeviceExternalSemaphoreProperties

vkGetPhysicalDeviceExternalSemaphoreProperties :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pExternalSemaphoreInfo" ::: Ptr VkPhysicalDeviceExternalSemaphoreInfo) -> ("pExternalSemaphoreProperties" ::: Ptr VkExternalSemaphoreProperties) -> IO ()

vkGetPhysicalDeviceExternalSemaphoreProperties - Function for querying external semaphore handle capabilities.

Parameters

physicalDevice is the physical device from which to query the semaphore capabilities.

pExternalSemaphoreInfo points to an instance of the VkPhysicalDeviceExternalSemaphoreInfo structure, describing the parameters that would be consumed by vkCreateSemaphore.

pExternalSemaphoreProperties points to an instance of the VkExternalSemaphoreProperties structure in which capabilities are returned.

Valid Usage (Implicit)

See Also

VkExternalSemaphoreProperties, VkPhysicalDevice, VkPhysicalDeviceExternalSemaphoreInfo

pattern VK_STRUCTURE_TYPE_EXTERNAL_SEMAPHORE_PROPERTIES :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_SEMAPHORE_INFO :: VkStructureType