Graphics.Vulkan.C.Core11.Promoted_from_VK_KHR_get_memory_requirements2

Documentation

data VkBufferMemoryRequirementsInfo2

VkBufferMemoryRequirementsInfo2 - (None)

Valid Usage (Implicit)

See Also

VkBuffer, VkStructureType, vkGetBufferMemoryRequirements2, vkGetBufferMemoryRequirements2KHR

Constructors

VkBufferMemoryRequirementsInfo2

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_BUFFER_MEMORY_REQUIREMENTS_INFO_2

vkPNext :: Ptr ()

pNext must be NULL

vkBuffer :: VkBuffer

buffer must be a valid VkBuffer handle

data VkImageMemoryRequirementsInfo2

VkImageMemoryRequirementsInfo2 - (None)

Valid Usage (Implicit)

See Also

VkImage, VkStructureType, vkGetImageMemoryRequirements2, vkGetImageMemoryRequirements2KHR

Constructors

VkImageMemoryRequirementsInfo2

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_IMAGE_MEMORY_REQUIREMENTS_INFO_2

vkPNext :: Ptr ()

pNext must be NULL or a pointer to a valid instance of VkImagePlaneMemoryRequirementsInfo

vkImage :: VkImage

image must be a valid VkImage handle

data VkImageSparseMemoryRequirementsInfo2

VkImageSparseMemoryRequirementsInfo2 - (None)

Valid Usage (Implicit)

See Also

VkImage, VkStructureType, vkGetImageSparseMemoryRequirements2, vkGetImageSparseMemoryRequirements2KHR

Constructors

VkImageSparseMemoryRequirementsInfo2

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_IMAGE_SPARSE_MEMORY_REQUIREMENTS_INFO_2

vkPNext :: Ptr ()

pNext must be NULL

vkImage :: VkImage

image must be a valid VkImage handle

data VkMemoryRequirements2

VkMemoryRequirements2 - Structure specifying memory requirements

Valid Usage (Implicit)

See Also

VkMemoryRequirements, VkStructureType, vkGetBufferMemoryRequirements2, vkGetBufferMemoryRequirements2KHR, vkGetImageMemoryRequirements2, vkGetImageMemoryRequirements2KHR

Constructors

VkMemoryRequirements2

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_MEMORY_REQUIREMENTS_2

vkPNext :: Ptr ()

pNext must be NULL or a pointer to a valid instance of VkMemoryDedicatedRequirements

vkMemoryRequirements :: VkMemoryRequirements

memoryRequirements is a structure of type VkMemoryRequirements describing the memory requirements of the resource.

type VkMemoryRequirements2KHR = VkMemoryRequirements2

pattern VkMemoryRequirements2KHR :: ("sType" ::: VkStructureType) -> ("pNext" ::: Ptr ()) -> ("memoryRequirements" ::: VkMemoryRequirements) -> VkMemoryRequirements2KHR

data VkSparseImageMemoryRequirements2

VkSparseImageMemoryRequirements2 - (None)

Valid Usage (Implicit)

See Also

VkSparseImageMemoryRequirements, VkStructureType, vkGetImageSparseMemoryRequirements2, vkGetImageSparseMemoryRequirements2KHR

Constructors

VkSparseImageMemoryRequirements2

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_SPARSE_IMAGE_MEMORY_REQUIREMENTS_2

vkPNext :: Ptr ()

pNext must be NULL

vkMemoryRequirements :: VkSparseImageMemoryRequirements

memoryRequirements is a structure of type VkSparseImageMemoryRequirements describing the memory requirements of the sparse image.

type FN_vkGetBufferMemoryRequirements2 = ("device" ::: VkDevice) -> ("pInfo" ::: Ptr VkBufferMemoryRequirementsInfo2) -> ("pMemoryRequirements" ::: Ptr VkMemoryRequirements2) -> IO ()

type PFN_vkGetBufferMemoryRequirements2 = FunPtr FN_vkGetBufferMemoryRequirements2

vkGetBufferMemoryRequirements2 :: DeviceCmds -> ("device" ::: VkDevice) -> ("pInfo" ::: Ptr VkBufferMemoryRequirementsInfo2) -> ("pMemoryRequirements" ::: Ptr VkMemoryRequirements2) -> IO ()

vkGetBufferMemoryRequirements2 - Returns the memory requirements for specified Vulkan object

Parameters

device is the logical device that owns the buffer.

pInfo is a pointer to an instance of the VkBufferMemoryRequirementsInfo2 structure containing parameters required for the memory requirements query.

pMemoryRequirements points to an instance of the VkMemoryRequirements2 structure in which the memory requirements of the buffer object are returned.

Valid Usage (Implicit)

See Also

VkBufferMemoryRequirementsInfo2, VkDevice, VkMemoryRequirements2

type FN_vkGetImageMemoryRequirements2 = ("device" ::: VkDevice) -> ("pInfo" ::: Ptr VkImageMemoryRequirementsInfo2) -> ("pMemoryRequirements" ::: Ptr VkMemoryRequirements2) -> IO ()

type PFN_vkGetImageMemoryRequirements2 = FunPtr FN_vkGetImageMemoryRequirements2

vkGetImageMemoryRequirements2 :: DeviceCmds -> ("device" ::: VkDevice) -> ("pInfo" ::: Ptr VkImageMemoryRequirementsInfo2) -> ("pMemoryRequirements" ::: Ptr VkMemoryRequirements2) -> IO ()

vkGetImageMemoryRequirements2 - Returns the memory requirements for specified Vulkan object

Parameters

device is the logical device that owns the image.

pInfo is a pointer to an instance of the VkImageMemoryRequirementsInfo2 structure containing parameters required for the memory requirements query.

pMemoryRequirements points to an instance of the VkMemoryRequirements2 structure in which the memory requirements of the image object are returned.

Valid Usage (Implicit)

See Also

VkDevice, VkImageMemoryRequirementsInfo2, VkMemoryRequirements2

type FN_vkGetImageSparseMemoryRequirements2 = ("device" ::: VkDevice) -> ("pInfo" ::: Ptr VkImageSparseMemoryRequirementsInfo2) -> ("pSparseMemoryRequirementCount" ::: Ptr Word32) -> ("pSparseMemoryRequirements" ::: Ptr VkSparseImageMemoryRequirements2) -> IO ()

type PFN_vkGetImageSparseMemoryRequirements2 = FunPtr FN_vkGetImageSparseMemoryRequirements2

vkGetImageSparseMemoryRequirements2 :: DeviceCmds -> ("device" ::: VkDevice) -> ("pInfo" ::: Ptr VkImageSparseMemoryRequirementsInfo2) -> ("pSparseMemoryRequirementCount" ::: Ptr Word32) -> ("pSparseMemoryRequirements" ::: Ptr VkSparseImageMemoryRequirements2) -> IO ()

vkGetImageSparseMemoryRequirements2 - Query the memory requirements for a sparse image

Parameters

device is the logical device that owns the image.

pInfo is a pointer to an instance of the VkImageSparseMemoryRequirementsInfo2 structure containing parameters required for the memory requirements query.

pSparseMemoryRequirementCount is a pointer to an integer related to the number of sparse memory requirements available or queried, as described below.

pSparseMemoryRequirements is either NULL or a pointer to an array of VkSparseImageMemoryRequirements2 structures.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pInfo must be a valid pointer to a valid VkImageSparseMemoryRequirementsInfo2 structure

pSparseMemoryRequirementCount must be a valid pointer to a uint32_t value

If the value referenced by pSparseMemoryRequirementCount is not 0, and pSparseMemoryRequirements is not NULL, pSparseMemoryRequirements must be a valid pointer to an array of pSparseMemoryRequirementCount VkSparseImageMemoryRequirements2 structures

See Also

VkDevice, VkImageSparseMemoryRequirementsInfo2, VkSparseImageMemoryRequirements2

pattern VK_STRUCTURE_TYPE_BUFFER_MEMORY_REQUIREMENTS_INFO_2 :: VkStructureType

pattern VK_STRUCTURE_TYPE_IMAGE_MEMORY_REQUIREMENTS_INFO_2 :: VkStructureType

pattern VK_STRUCTURE_TYPE_IMAGE_SPARSE_MEMORY_REQUIREMENTS_INFO_2 :: VkStructureType

pattern VK_STRUCTURE_TYPE_MEMORY_REQUIREMENTS_2 :: VkStructureType

pattern VK_STRUCTURE_TYPE_SPARSE_IMAGE_MEMORY_REQUIREMENTS_2 :: VkStructureType