Graphics.Vulkan.C.Core11.Promoted_from_VK_KHR_get_physical_device_properties2

Documentation

data VkFormatProperties2

VkFormatProperties2 - Structure specifying image format properties

Valid Usage (Implicit)

See Also

VkFormatProperties, VkStructureType, vkGetPhysicalDeviceFormatProperties2, vkGetPhysicalDeviceFormatProperties2KHR

Constructors

VkFormatProperties2

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_FORMAT_PROPERTIES_2

vkPNext :: Ptr ()

pNext must be NULL or a pointer to a valid instance of VkDrmFormatModifierPropertiesListEXT

vkFormatProperties :: VkFormatProperties

formatProperties is a structure of type VkFormatProperties describing features supported by the requested format.

data VkImageFormatProperties2

VkImageFormatProperties2 - Structure specifying an image format properties

Description

If the combination of parameters to vkGetPhysicalDeviceImageFormatProperties2 is not supported by the implementation for use in vkCreateImage, then all members of imageFormatProperties will be filled with zero.

Note

Filling imageFormatProperties with zero for unsupported formats is an exception to the usual rule that output structures have undefined contents on error. This exception was unintentional, but is preserved for backwards compatibility. This exeption only applies to imageFormatProperties, not sType, pNext, or any structures chained from pNext.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMAGE_FORMAT_PROPERTIES_2

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkAndroidHardwareBufferUsageANDROID, VkExternalImageFormatProperties, VkFilterCubicImageViewImageFormatPropertiesEXT, VkSamplerYcbcrConversionImageFormatProperties, or VkTextureLODGatherFormatPropertiesAMD

Each sType member in the pNext chain must be unique

See Also

VkImageFormatProperties, VkStructureType, vkGetPhysicalDeviceImageFormatProperties2, vkGetPhysicalDeviceImageFormatProperties2KHR

Constructors

VkImageFormatProperties2

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure. The pNext chain of VkImageFormatProperties2 is used to allow the specification of additional capabilities to be returned from vkGetPhysicalDeviceImageFormatProperties2.

vkImageFormatProperties :: VkImageFormatProperties

imageFormatProperties is an instance of a VkImageFormatProperties structure in which capabilities are returned.

data VkPhysicalDeviceFeatures2

VkPhysicalDeviceFeatures2 - Structure describing the fine-grained features that can be supported by an implementation

Members

The VkPhysicalDeviceFeatures2 structure is defined as:

Description

The pNext chain of this structure is used to extend the structure with features defined by extensions. This structure can be used in vkGetPhysicalDeviceFeatures2 or can be in the pNext chain of a VkDeviceCreateInfo structure, in which case it controls which features are enabled in the device in lieu of pEnabledFeatures.

Valid Usage (Implicit)

See Also

VkPhysicalDeviceFeatures, VkStructureType, vkGetPhysicalDeviceFeatures2, vkGetPhysicalDeviceFeatures2KHR

Constructors

VkPhysicalDeviceFeatures2

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FEATURES_2

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkFeatures :: VkPhysicalDeviceFeatures

features is a structure of type VkPhysicalDeviceFeatures describing the fine-grained features of the Vulkan 1.0 API.

data VkPhysicalDeviceImageFormatInfo2

VkPhysicalDeviceImageFormatInfo2 - Structure specifying image creation parameters

Description

The members of VkPhysicalDeviceImageFormatInfo2 correspond to the arguments to vkGetPhysicalDeviceImageFormatProperties, with sType and pNext added for extensibility.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGE_FORMAT_INFO_2

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkImageFormatListCreateInfoKHR, VkImageStencilUsageCreateInfoEXT, VkPhysicalDeviceExternalImageFormatInfo, VkPhysicalDeviceImageDrmFormatModifierInfoEXT, or VkPhysicalDeviceImageViewImageFormatInfoEXT

Each sType member in the pNext chain must be unique

format must be a valid VkFormat value

type must be a valid VkImageType value

tiling must be a valid VkImageTiling value

usage must be a valid combination of VkImageUsageFlagBits values

usage must not be 0

flags must be a valid combination of VkImageCreateFlagBits values

See Also

VkFormat, VkImageCreateFlags, VkImageTiling, VkImageType, VkImageUsageFlags, VkStructureType, vkGetPhysicalDeviceImageFormatProperties2, vkGetPhysicalDeviceImageFormatProperties2KHR

Constructors

VkPhysicalDeviceImageFormatInfo2

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure. The pNext chain of VkPhysicalDeviceImageFormatInfo2 is used to provide additional image parameters to vkGetPhysicalDeviceImageFormatProperties2.

vkFormat :: VkFormat

format is a VkFormat value indicating the image format, corresponding to VkImageCreateInfo::format.

vkType :: VkImageType

type is a VkImageType value indicating the image type, corresponding to VkImageCreateInfo::imageType.

vkTiling :: VkImageTiling

tiling is a VkImageTiling value indicating the image tiling, corresponding to VkImageCreateInfo::tiling.

vkUsage :: VkImageUsageFlags

usage is a bitmask of VkImageUsageFlagBits indicating the intended usage of the image, corresponding to VkImageCreateInfo::usage.

vkFlags :: VkImageCreateFlags

flags is a bitmask of VkImageCreateFlagBits indicating additional parameters of the image, corresponding to VkImageCreateInfo::flags.

data VkPhysicalDeviceMemoryProperties2

VkPhysicalDeviceMemoryProperties2 - Structure specifying physical device memory properties

Valid Usage (Implicit)

See Also

VkPhysicalDeviceMemoryProperties, VkStructureType, vkGetPhysicalDeviceMemoryProperties2, vkGetPhysicalDeviceMemoryProperties2KHR

Constructors

VkPhysicalDeviceMemoryProperties2

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MEMORY_PROPERTIES_2

vkPNext :: Ptr ()

pNext must be NULL or a pointer to a valid instance of VkPhysicalDeviceMemoryBudgetPropertiesEXT

vkMemoryProperties :: VkPhysicalDeviceMemoryProperties

memoryProperties is a structure of type VkPhysicalDeviceMemoryProperties which is populated with the same values as in vkGetPhysicalDeviceMemoryProperties.

data VkPhysicalDeviceProperties2

VkPhysicalDeviceProperties2 - Structure specifying physical device properties

Description

The pNext chain of this structure is used to extend the structure with properties defined by extensions.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PROPERTIES_2

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkPhysicalDeviceBlendOperationAdvancedPropertiesEXT, VkPhysicalDeviceConservativeRasterizationPropertiesEXT, VkPhysicalDeviceCooperativeMatrixPropertiesNV, VkPhysicalDeviceDepthStencilResolvePropertiesKHR, VkPhysicalDeviceDescriptorIndexingPropertiesEXT, VkPhysicalDeviceDiscardRectanglePropertiesEXT, VkPhysicalDeviceDriverPropertiesKHR, VkPhysicalDeviceExternalMemoryHostPropertiesEXT, VkPhysicalDeviceFloatControlsPropertiesKHR, VkPhysicalDeviceFragmentDensityMapPropertiesEXT, VkPhysicalDeviceIDProperties, VkPhysicalDeviceInlineUniformBlockPropertiesEXT, VkPhysicalDeviceMaintenance3Properties, VkPhysicalDeviceMeshShaderPropertiesNV, VkPhysicalDeviceMultiviewPerViewAttributesPropertiesNVX, VkPhysicalDeviceMultiviewProperties, VkPhysicalDevicePCIBusInfoPropertiesEXT, VkPhysicalDevicePointClippingProperties, VkPhysicalDeviceProtectedMemoryProperties, VkPhysicalDevicePushDescriptorPropertiesKHR, VkPhysicalDeviceRayTracingPropertiesNV, VkPhysicalDeviceSampleLocationsPropertiesEXT, VkPhysicalDeviceSamplerFilterMinmaxPropertiesEXT, VkPhysicalDeviceShaderCorePropertiesAMD, VkPhysicalDeviceShadingRateImagePropertiesNV, VkPhysicalDeviceSubgroupProperties, VkPhysicalDeviceTransformFeedbackPropertiesEXT, or VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT

Each sType member in the pNext chain must be unique

See Also

VkPhysicalDeviceProperties, VkStructureType, vkGetPhysicalDeviceProperties2, vkGetPhysicalDeviceProperties2KHR

Constructors

VkPhysicalDeviceProperties2

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkProperties :: VkPhysicalDeviceProperties

properties is a structure of type VkPhysicalDeviceProperties describing the properties of the physical device. This structure is written with the same values as if it were written by vkGetPhysicalDeviceProperties.

data VkPhysicalDeviceSparseImageFormatInfo2

VkPhysicalDeviceSparseImageFormatInfo2 - Structure specifying sparse image format inputs

Valid Usage (Implicit)

See Also

VkFormat, VkImageTiling, VkImageType, VkImageUsageFlags, VkSampleCountFlagBits, VkStructureType, vkGetPhysicalDeviceSparseImageFormatProperties2, vkGetPhysicalDeviceSparseImageFormatProperties2KHR

Constructors

VkPhysicalDeviceSparseImageFormatInfo2

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SPARSE_IMAGE_FORMAT_INFO_2

vkPNext :: Ptr ()

pNext must be NULL

vkFormat :: VkFormat

format must be a valid VkFormat value

vkType :: VkImageType

type must be a valid VkImageType value

vkSamples :: VkSampleCountFlagBits

samples must be a valid VkSampleCountFlagBits value

vkUsage :: VkImageUsageFlags

usage must not be 0

vkTiling :: VkImageTiling

tiling must be a valid VkImageTiling value

data VkQueueFamilyProperties2

VkQueueFamilyProperties2 - Structure providing information about a queue family

Valid Usage (Implicit)

See Also

VkQueueFamilyProperties, VkStructureType, vkGetPhysicalDeviceQueueFamilyProperties2, vkGetPhysicalDeviceQueueFamilyProperties2KHR

Constructors

VkQueueFamilyProperties2

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_QUEUE_FAMILY_PROPERTIES_2

vkPNext :: Ptr ()

pNext must be NULL or a pointer to a valid instance of VkQueueFamilyCheckpointPropertiesNV

vkQueueFamilyProperties :: VkQueueFamilyProperties

queueFamilyProperties is a structure of type VkQueueFamilyProperties which is populated with the same values as in vkGetPhysicalDeviceQueueFamilyProperties.

data VkSparseImageFormatProperties2

VkSparseImageFormatProperties2 - Structure specifying sparse image format properties

Valid Usage (Implicit)

See Also

VkSparseImageFormatProperties, VkStructureType, vkGetPhysicalDeviceSparseImageFormatProperties2, vkGetPhysicalDeviceSparseImageFormatProperties2KHR

Constructors

VkSparseImageFormatProperties2

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_SPARSE_IMAGE_FORMAT_PROPERTIES_2

vkPNext :: Ptr ()

pNext must be NULL

vkProperties :: VkSparseImageFormatProperties

properties is a structure of type VkSparseImageFormatProperties which is populated with the same values as in vkGetPhysicalDeviceSparseImageFormatProperties.

type FN_vkGetPhysicalDeviceFeatures2 = ("physicalDevice" ::: VkPhysicalDevice) -> ("pFeatures" ::: Ptr VkPhysicalDeviceFeatures2) -> IO ()

type PFN_vkGetPhysicalDeviceFeatures2 = FunPtr FN_vkGetPhysicalDeviceFeatures2

vkGetPhysicalDeviceFeatures2 :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pFeatures" ::: Ptr VkPhysicalDeviceFeatures2) -> IO ()

vkGetPhysicalDeviceFeatures2 - Reports capabilities of a physical device

Parameters

physicalDevice is the physical device from which to query the supported features.

pFeatures is a pointer to a VkPhysicalDeviceFeatures2 structure in which the physical device features are returned.

Description

Each structure in pFeatures and its pNext chain contain members corresponding to fine-grained features. vkGetPhysicalDeviceFeatures2 writes each member to a boolean value indicating whether that feature is supported.

Valid Usage (Implicit)

See Also

VkPhysicalDevice, VkPhysicalDeviceFeatures2

type FN_vkGetPhysicalDeviceFormatProperties2 = ("physicalDevice" ::: VkPhysicalDevice) -> ("format" ::: VkFormat) -> ("pFormatProperties" ::: Ptr VkFormatProperties2) -> IO ()

type PFN_vkGetPhysicalDeviceFormatProperties2 = FunPtr FN_vkGetPhysicalDeviceFormatProperties2

vkGetPhysicalDeviceFormatProperties2 :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("format" ::: VkFormat) -> ("pFormatProperties" ::: Ptr VkFormatProperties2) -> IO ()

vkGetPhysicalDeviceFormatProperties2 - Lists physical device’s format capabilities

Parameters

physicalDevice is the physical device from which to query the format properties.

format is the format whose properties are queried.

pFormatProperties is a pointer to a VkFormatProperties2 structure in which physical device properties for format are returned.

Description

vkGetPhysicalDeviceFormatProperties2 behaves similarly to vkGetPhysicalDeviceFormatProperties, with the ability to return extended information in a pNext chain of output structures.

Valid Usage (Implicit)

See Also

VkFormat, VkFormatProperties2, VkPhysicalDevice

type FN_vkGetPhysicalDeviceImageFormatProperties2 = ("physicalDevice" ::: VkPhysicalDevice) -> ("pImageFormatInfo" ::: Ptr VkPhysicalDeviceImageFormatInfo2) -> ("pImageFormatProperties" ::: Ptr VkImageFormatProperties2) -> IO VkResult

type PFN_vkGetPhysicalDeviceImageFormatProperties2 = FunPtr FN_vkGetPhysicalDeviceImageFormatProperties2

vkGetPhysicalDeviceImageFormatProperties2 :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pImageFormatInfo" ::: Ptr VkPhysicalDeviceImageFormatInfo2) -> ("pImageFormatProperties" ::: Ptr VkImageFormatProperties2) -> IO VkResult

vkGetPhysicalDeviceImageFormatProperties2 - Lists physical device’s image format capabilities

Parameters

physicalDevice is the physical device from which to query the image capabilities.

pImageFormatInfo points to an instance of the VkPhysicalDeviceImageFormatInfo2 structure, describing the parameters that would be consumed by vkCreateImage.

pImageFormatProperties points to an instance of the VkImageFormatProperties2 structure in which capabilities are returned.

Description

vkGetPhysicalDeviceImageFormatProperties2 behaves similarly to vkGetPhysicalDeviceImageFormatProperties, with the ability to return extended information in a pNext chain of output structures.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_FORMAT_NOT_SUPPORTED

See Also

VkImageFormatProperties2, VkPhysicalDevice, VkPhysicalDeviceImageFormatInfo2

type FN_vkGetPhysicalDeviceMemoryProperties2 = ("physicalDevice" ::: VkPhysicalDevice) -> ("pMemoryProperties" ::: Ptr VkPhysicalDeviceMemoryProperties2) -> IO ()

type PFN_vkGetPhysicalDeviceMemoryProperties2 = FunPtr FN_vkGetPhysicalDeviceMemoryProperties2

vkGetPhysicalDeviceMemoryProperties2 :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pMemoryProperties" ::: Ptr VkPhysicalDeviceMemoryProperties2) -> IO ()

vkGetPhysicalDeviceMemoryProperties2 - Reports memory information for the specified physical device

Parameters

physicalDevice is the handle to the device to query.

pMemoryProperties points to an instance of VkPhysicalDeviceMemoryProperties2 structure in which the properties are returned.

Description

vkGetPhysicalDeviceMemoryProperties2 behaves similarly to vkGetPhysicalDeviceMemoryProperties, with the ability to return extended information in a pNext chain of output structures.

Valid Usage (Implicit)

See Also

VkPhysicalDevice, VkPhysicalDeviceMemoryProperties2

type FN_vkGetPhysicalDeviceProperties2 = ("physicalDevice" ::: VkPhysicalDevice) -> ("pProperties" ::: Ptr VkPhysicalDeviceProperties2) -> IO ()

type PFN_vkGetPhysicalDeviceProperties2 = FunPtr FN_vkGetPhysicalDeviceProperties2

vkGetPhysicalDeviceProperties2 :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pProperties" ::: Ptr VkPhysicalDeviceProperties2) -> IO ()

vkGetPhysicalDeviceProperties2 - Returns properties of a physical device

Parameters

physicalDevice is the handle to the physical device whose properties will be queried.

pProperties points to an instance of the VkPhysicalDeviceProperties2 structure, that will be filled with returned information.

Description

Each structure in pProperties and its pNext chain contain members corresponding to properties or implementation-dependent limits. vkGetPhysicalDeviceProperties2 writes each member to a value indicating the value of that property or limit.

Valid Usage (Implicit)

See Also

VkPhysicalDevice, VkPhysicalDeviceProperties2

type FN_vkGetPhysicalDeviceQueueFamilyProperties2 = ("physicalDevice" ::: VkPhysicalDevice) -> ("pQueueFamilyPropertyCount" ::: Ptr Word32) -> ("pQueueFamilyProperties" ::: Ptr VkQueueFamilyProperties2) -> IO ()

type PFN_vkGetPhysicalDeviceQueueFamilyProperties2 = FunPtr FN_vkGetPhysicalDeviceQueueFamilyProperties2

vkGetPhysicalDeviceQueueFamilyProperties2 :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pQueueFamilyPropertyCount" ::: Ptr Word32) -> ("pQueueFamilyProperties" ::: Ptr VkQueueFamilyProperties2) -> IO ()

vkGetPhysicalDeviceQueueFamilyProperties2 - Reports properties of the queues of the specified physical device

Parameters

physicalDevice is the handle to the physical device whose properties will be queried.

pQueueFamilyPropertyCount is a pointer to an integer related to the number of queue families available or queried, as described in vkGetPhysicalDeviceQueueFamilyProperties.

pQueueFamilyProperties is either NULL or a pointer to an array of VkQueueFamilyProperties2 structures.

Description

vkGetPhysicalDeviceQueueFamilyProperties2 behaves similarly to vkGetPhysicalDeviceQueueFamilyProperties, with the ability to return extended information in a pNext chain of output structures.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pQueueFamilyPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pQueueFamilyPropertyCount is not 0, and pQueueFamilyProperties is not NULL, pQueueFamilyProperties must be a valid pointer to an array of pQueueFamilyPropertyCount VkQueueFamilyProperties2 structures

See Also

VkPhysicalDevice, VkQueueFamilyProperties2

type FN_vkGetPhysicalDeviceSparseImageFormatProperties2 = ("physicalDevice" ::: VkPhysicalDevice) -> ("pFormatInfo" ::: Ptr VkPhysicalDeviceSparseImageFormatInfo2) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkSparseImageFormatProperties2) -> IO ()

type PFN_vkGetPhysicalDeviceSparseImageFormatProperties2 = FunPtr FN_vkGetPhysicalDeviceSparseImageFormatProperties2

vkGetPhysicalDeviceSparseImageFormatProperties2 :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pFormatInfo" ::: Ptr VkPhysicalDeviceSparseImageFormatInfo2) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkSparseImageFormatProperties2) -> IO ()

vkGetPhysicalDeviceSparseImageFormatProperties2 - Retrieve properties of an image format applied to sparse images

Parameters

physicalDevice is the physical device from which to query the sparse image capabilities.

pFormatInfo is a pointer to a structure of type VkPhysicalDeviceSparseImageFormatInfo2 containing input parameters to the command.

pPropertyCount is a pointer to an integer related to the number of sparse format properties available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkSparseImageFormatProperties2 structures.

Description

vkGetPhysicalDeviceSparseImageFormatProperties2 behaves identically to vkGetPhysicalDeviceSparseImageFormatProperties, with the ability to return extended information by adding extension structures to the pNext chain of its pProperties parameter.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pFormatInfo must be a valid pointer to a valid VkPhysicalDeviceSparseImageFormatInfo2 structure

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkSparseImageFormatProperties2 structures

See Also

VkPhysicalDevice, VkPhysicalDeviceSparseImageFormatInfo2, VkSparseImageFormatProperties2

pattern VK_STRUCTURE_TYPE_FORMAT_PROPERTIES_2 :: VkStructureType

pattern VK_STRUCTURE_TYPE_IMAGE_FORMAT_PROPERTIES_2 :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_FEATURES_2 :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGE_FORMAT_INFO_2 :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MEMORY_PROPERTIES_2 :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PROPERTIES_2 :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SPARSE_IMAGE_FORMAT_INFO_2 :: VkStructureType

pattern VK_STRUCTURE_TYPE_QUEUE_FAMILY_PROPERTIES_2 :: VkStructureType

pattern VK_STRUCTURE_TYPE_SPARSE_IMAGE_FORMAT_PROPERTIES_2 :: VkStructureType