Graphics.Vulkan.C.Core11.Promoted_from_VK_KHR_maintenance2

Documentation

data VkImageViewUsageCreateInfo

VkImageViewUsageCreateInfo - Specify the intended usage of an image view

Description

When this structure is chained to VkImageViewCreateInfo the usage field overrides the implicit usage parameter inherited from image creation time and its value is used instead for the purposes of determining the valid usage conditions of VkImageViewCreateInfo.

Valid Usage (Implicit)

See Also

VkImageUsageFlags, VkStructureType

Constructors

VkImageViewUsageCreateInfo

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_IMAGE_VIEW_USAGE_CREATE_INFO

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkUsage :: VkImageUsageFlags

usage must not be 0

data VkInputAttachmentAspectReference

VkInputAttachmentAspectReference - Structure specifying a subpass/input attachment pair and an aspect mask that can be read.

Valid Usage (Implicit)

See Also

VkImageAspectFlags, VkRenderPassInputAttachmentAspectCreateInfo

Constructors

VkInputAttachmentAspectReference

Fields

vkSubpass :: Word32

subpass is an index into the pSubpasses array of the parent VkRenderPassCreateInfo structure.

vkInputAttachmentIndex :: Word32

inputAttachmentIndex is an index into the pInputAttachments of the specified subpass.

vkAspectMask :: VkImageAspectFlags

aspectMask must not be 0

data VkPhysicalDevicePointClippingProperties

VkPhysicalDevicePointClippingProperties - Structure describing the point clipping behavior supported by an implementation

Members

The members of the VkPhysicalDevicePointClippingProperties structure describe the following implementation-dependent limit:

Description

If the VkPhysicalDevicePointClippingProperties structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkPointClippingBehavior, VkStructureType

Constructors

VkPhysicalDevicePointClippingProperties

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_POINT_CLIPPING_PROPERTIES

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkPointClippingBehavior :: VkPointClippingBehavior

pointClippingBehavior is the point clipping behavior supported by the implementation, and is of type VkPointClippingBehavior.

data VkPipelineTessellationDomainOriginStateCreateInfo

VkPipelineTessellationDomainOriginStateCreateInfo - Structure specifying the orientation of the tessellation domain

Description

If the VkPipelineTessellationDomainOriginStateCreateInfo structure is included in the pNext chain of VkPipelineTessellationStateCreateInfo, it controls the origin of the tessellation domain. If this structure is not present, it is as if domainOrigin were VK_TESSELLATION_DOMAIN_ORIGIN_UPPER_LEFT.

Valid Usage (Implicit)

See Also

VkStructureType, VkTessellationDomainOrigin

Constructors

VkPipelineTessellationDomainOriginStateCreateInfo

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PIPELINE_TESSELLATION_DOMAIN_ORIGIN_STATE_CREATE_INFO

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkDomainOrigin :: VkTessellationDomainOrigin

domainOrigin must be a valid VkTessellationDomainOrigin value

newtype VkPointClippingBehavior

VkPointClippingBehavior - Enum specifying the point clipping behavior

See Also

VkPhysicalDevicePointClippingProperties

Constructors

VkPointClippingBehavior Int32

pattern VK_POINT_CLIPPING_BEHAVIOR_ALL_CLIP_PLANES :: VkPointClippingBehavior

VK_POINT_CLIPPING_BEHAVIOR_ALL_CLIP_PLANES specifies that the primitive is discarded if the vertex lies outside any clip plane, including the planes bounding the view volume.

pattern VK_POINT_CLIPPING_BEHAVIOR_USER_CLIP_PLANES_ONLY :: VkPointClippingBehavior

VK_POINT_CLIPPING_BEHAVIOR_USER_CLIP_PLANES_ONLY specifies that the primitive is discarded only if the vertex lies outside any user clip plane.

data VkRenderPassInputAttachmentAspectCreateInfo

VkRenderPassInputAttachmentAspectCreateInfo - Structure specifying, for a given subpass/input attachment pair, which aspect can be read.

Valid Usage (Implicit)

See Also

VkInputAttachmentAspectReference, VkStructureType

Constructors

VkRenderPassInputAttachmentAspectCreateInfo

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_RENDER_PASS_INPUT_ATTACHMENT_ASPECT_CREATE_INFO

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkAspectReferenceCount :: Word32

aspectReferenceCount must be greater than 0

vkPAspectReferences :: Ptr VkInputAttachmentAspectReference

pAspectReferences must be a valid pointer to an array of aspectReferenceCount valid VkInputAttachmentAspectReference structures

newtype VkTessellationDomainOrigin

VkTessellationDomainOrigin - Enum describing tessellation domain origin

Description

This enum affects how the VertexOrderCw and VertexOrderCcw tessellation execution modes are interpreted, since the winding is defined relative to the orientation of the domain.

See Also

VkPipelineTessellationDomainOriginStateCreateInfo

Constructors

VkTessellationDomainOrigin Int32

pattern VK_TESSELLATION_DOMAIN_ORIGIN_UPPER_LEFT :: VkTessellationDomainOrigin

VK_TESSELLATION_DOMAIN_ORIGIN_UPPER_LEFT specifies that the origin of the domain space is in the upper left corner, as shown in figure https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#img-tessellation-topology-ul.

pattern VK_TESSELLATION_DOMAIN_ORIGIN_LOWER_LEFT :: VkTessellationDomainOrigin

VK_TESSELLATION_DOMAIN_ORIGIN_LOWER_LEFT specifies that the origin of the domain space is in the lower left corner, as shown in figure https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#img-tessellation-topology-ll.

pattern VK_IMAGE_CREATE_BLOCK_TEXEL_VIEW_COMPATIBLE_BIT :: VkImageCreateFlagBits

pattern VK_IMAGE_CREATE_EXTENDED_USAGE_BIT :: VkImageCreateFlagBits

pattern VK_IMAGE_LAYOUT_DEPTH_ATTACHMENT_STENCIL_READ_ONLY_OPTIMAL :: VkImageLayout

pattern VK_IMAGE_LAYOUT_DEPTH_READ_ONLY_STENCIL_ATTACHMENT_OPTIMAL :: VkImageLayout

pattern VK_STRUCTURE_TYPE_IMAGE_VIEW_USAGE_CREATE_INFO :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_POINT_CLIPPING_PROPERTIES :: VkStructureType

pattern VK_STRUCTURE_TYPE_PIPELINE_TESSELLATION_DOMAIN_ORIGIN_STATE_CREATE_INFO :: VkStructureType

pattern VK_STRUCTURE_TYPE_RENDER_PASS_INPUT_ATTACHMENT_ASPECT_CREATE_INFO :: VkStructureType