Graphics.Vulkan.C.Core11.Promoted_from_VK_KHR_maintenance3

Documentation

data VkDescriptorSetLayoutSupport

VkDescriptorSetLayoutSupport - Structure returning information about whether a descriptor set layout can be supported

Description

supported is set to VK_TRUE if the descriptor set can be created, or else is set to VK_FALSE.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType, vkGetDescriptorSetLayoutSupport, vkGetDescriptorSetLayoutSupportKHR

Constructors

VkDescriptorSetLayoutSupport

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_SUPPORT

vkPNext :: Ptr ()

pNext must be NULL or a pointer to a valid instance of VkDescriptorSetVariableDescriptorCountLayoutSupportEXT

vkSupported :: VkBool32

supported specifies whether the descriptor set layout can be created.

data VkPhysicalDeviceMaintenance3Properties

VkPhysicalDeviceMaintenance3Properties - Structure describing descriptor set properties

Members

The members of the VkPhysicalDeviceMaintenance3Properties structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceMaintenance3Properties structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkDeviceSize, VkStructureType

Constructors

VkPhysicalDeviceMaintenance3Properties

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MAINTENANCE_3_PROPERTIES

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkMaxPerSetDescriptors :: Word32

maxPerSetDescriptors is a maximum number of descriptors (summed over all descriptor types) in a single descriptor set that is guaranteed to satisfy any implementation-dependent constraints on the size of a descriptor set itself. Applications can query whether a descriptor set that goes beyond this limit is supported using vkGetDescriptorSetLayoutSupport.

vkMaxMemoryAllocationSize :: VkDeviceSize

maxMemoryAllocationSize is the maximum size of a memory allocation that can be created, even if there is more space available in the heap.

type FN_vkGetDescriptorSetLayoutSupport = ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkDescriptorSetLayoutCreateInfo) -> ("pSupport" ::: Ptr VkDescriptorSetLayoutSupport) -> IO ()

type PFN_vkGetDescriptorSetLayoutSupport = FunPtr FN_vkGetDescriptorSetLayoutSupport

vkGetDescriptorSetLayoutSupport :: DeviceCmds -> ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkDescriptorSetLayoutCreateInfo) -> ("pSupport" ::: Ptr VkDescriptorSetLayoutSupport) -> IO ()

vkGetDescriptorSetLayoutSupport - Query whether a descriptor set layout can be created

Parameters

device is the logical device that would create the descriptor set layout.

pCreateInfo is a pointer to an instance of the VkDescriptorSetLayoutCreateInfo structure specifying the state of the descriptor set layout object.

pSupport points to a VkDescriptorSetLayoutSupport structure in which information about support for the descriptor set layout object is returned.

Description

Some implementations have limitations on what fits in a descriptor set which are not easily expressible in terms of existing limits like maxDescriptorSet*, for example if all descriptor types share a limited space in memory but each descriptor is a different size or alignment. This command returns information about whether a descriptor set satisfies this limit. If the descriptor set layout satisfies the VkPhysicalDeviceMaintenance3Properties::maxPerSetDescriptors limit, this command is guaranteed to return VK_TRUE in VkDescriptorSetLayoutSupport::supported. If the descriptor set layout exceeds the VkPhysicalDeviceMaintenance3Properties::maxPerSetDescriptors limit, whether the descriptor set layout is supported is implementation-dependent and may depend on whether the descriptor sizes and alignments cause the layout to exceed an internal limit.

This command does not consider other limits such as maxPerStageDescriptor*, and so a descriptor set layout that is supported according to this command must still satisfy the pipeline layout limits such as maxPerStageDescriptor* in order to be used in a pipeline layout.

Note

This is a VkDevice query rather than VkPhysicalDevice because the answer may depend on enabled features.

Valid Usage (Implicit)

See Also

VkDescriptorSetLayoutCreateInfo, VkDescriptorSetLayoutSupport, VkDevice

pattern VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_SUPPORT :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MAINTENANCE_3_PROPERTIES :: VkStructureType