Graphics.Vulkan.C.Extensions.VK_AMD_display_native_hdr

Documentation

data VkDisplayNativeHdrSurfaceCapabilitiesAMD

VkDisplayNativeHdrSurfaceCapabilitiesAMD - Structure describing display native HDR specific capabilities of a surface

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

VkDisplayNativeHdrSurfaceCapabilitiesAMD

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DISPLAY_NATIVE_HDR_SURFACE_CAPABILITIES_AMD

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkLocalDimmingSupport :: VkBool32

localDimmingSupport specifies whether the surface supports local dimming. If this is VK_TRUE, VkSwapchainDisplayNativeHdrCreateInfoAMD can be used to explicitly enable or disable local dimming for the surface. Local dimming may also be overriden by vkSetLocalDimmingAMD during the lifetime of the swapchain.

data VkSwapchainDisplayNativeHdrCreateInfoAMD

VkSwapchainDisplayNativeHdrCreateInfoAMD - Structure specifying display native HDR parameters of a newly created swapchain object

Description

If the pNext chain of VkSwapchainCreateInfoKHR does not contain this structure, the default value for localDimmingEnable is VK_TRUE, meaning local dimming is initially enabled for the swapchain.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_SWAPCHAIN_DISPLAY_NATIVE_HDR_CREATE_INFO_AMD

Valid Usage

It is only valid to set localDimmingEnable to VK_TRUE if VkDisplayNativeHdrSurfaceCapabilitiesAMD::localDimmingSupport is supported.

See Also

VkBool32, VkStructureType

Constructors

VkSwapchainDisplayNativeHdrCreateInfoAMD

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkLocalDimmingEnable :: VkBool32

localDimmingEnable specifies whether local dimming is enabled for the swapchain.

type FN_vkSetLocalDimmingAMD = ("device" ::: VkDevice) -> ("swapChain" ::: VkSwapchainKHR) -> ("localDimmingEnable" ::: VkBool32) -> IO ()

type PFN_vkSetLocalDimmingAMD = FunPtr FN_vkSetLocalDimmingAMD

vkSetLocalDimmingAMD :: DeviceCmds -> ("device" ::: VkDevice) -> ("swapChain" ::: VkSwapchainKHR) -> ("localDimmingEnable" ::: VkBool32) -> IO ()

vkSetLocalDimmingAMD - Set Local Dimming

Parameters

device is the device associated with swapChain.

swapChain handle to enable local dimming.

localDimmingEnable specifies whether local dimming is enabled for the swapchain.

Valid Usage (Implicit)

device must be a valid VkDevice handle

swapChain must be a valid VkSwapchainKHR handle

Both of device, and swapChain must have been created, allocated, or retrieved from the same VkInstance

Valid Usage

It is only valid to call vkSetLocalDimmingAMD if VkDisplayNativeHdrSurfaceCapabilitiesAMD::localDimmingSupport is supported.

See Also

VkBool32, VkDevice, VkSwapchainKHR

pattern VK_AMD_DISPLAY_NATIVE_HDR_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_AMD_DISPLAY_NATIVE_HDR_SPEC_VERSION :: Integral a => a

pattern VK_COLOR_SPACE_DISPLAY_NATIVE_AMD :: VkColorSpaceKHR

pattern VK_STRUCTURE_TYPE_DISPLAY_NATIVE_HDR_SURFACE_CAPABILITIES_AMD :: VkStructureType

pattern VK_STRUCTURE_TYPE_SWAPCHAIN_DISPLAY_NATIVE_HDR_CREATE_INFO_AMD :: VkStructureType