Graphics.Vulkan.C.Extensions.VK_AMD_memory_overallocation_behavior

Documentation

data VkDeviceMemoryOverallocationCreateInfoAMD

VkDeviceMemoryOverallocationCreateInfoAMD - Specify memory overallocation behavior for a Vulkan device

Valid Usage (Implicit)

See Also

VkMemoryOverallocationBehaviorAMD, VkStructureType

Constructors

VkDeviceMemoryOverallocationCreateInfoAMD

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DEVICE_MEMORY_OVERALLOCATION_CREATE_INFO_AMD

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkOverallocationBehavior :: VkMemoryOverallocationBehaviorAMD

overallocationBehavior must be a valid VkMemoryOverallocationBehaviorAMD value

newtype VkMemoryOverallocationBehaviorAMD

VkMemoryOverallocationBehaviorAMD - Specify memory overallocation behavior

See Also

VkDeviceMemoryOverallocationCreateInfoAMD

Constructors

VkMemoryOverallocationBehaviorAMD Int32

pattern VK_MEMORY_OVERALLOCATION_BEHAVIOR_DEFAULT_AMD :: VkMemoryOverallocationBehaviorAMD

VK_MEMORY_OVERALLOCATION_BEHAVIOR_DEFAULT_AMD lets the implementation decide if overallocation should be allowed.

pattern VK_MEMORY_OVERALLOCATION_BEHAVIOR_ALLOWED_AMD :: VkMemoryOverallocationBehaviorAMD

VK_MEMORY_OVERALLOCATION_BEHAVIOR_ALLOWED_AMD specifies overallocation is allowed if platform permits.

pattern VK_MEMORY_OVERALLOCATION_BEHAVIOR_DISALLOWED_AMD :: VkMemoryOverallocationBehaviorAMD

VK_MEMORY_OVERALLOCATION_BEHAVIOR_DISALLOWED_AMD specifies the application is not allowed to allocate device memory beyond the heap sizes reported by VkPhysicalDeviceMemoryProperties. Allocations that are not explicitly made by the application within the scope of the Vulkan instance are not accounted for.

pattern VK_AMD_MEMORY_OVERALLOCATION_BEHAVIOR_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_AMD_MEMORY_OVERALLOCATION_BEHAVIOR_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_DEVICE_MEMORY_OVERALLOCATION_CREATE_INFO_AMD :: VkStructureType