Graphics.Vulkan.C.Extensions.VK_AMD_rasterization_order

Documentation

data VkPipelineRasterizationStateRasterizationOrderAMD

VkPipelineRasterizationStateRasterizationOrderAMD - Structure defining rasterization order for a graphics pipeline

Valid Usage (Implicit)

If the https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#VK_AMD_rasterization_order device extension is not enabled or the application does not request a particular rasterization order through specifying a VkPipelineRasterizationStateRasterizationOrderAMD structure then the rasterization order used by the graphics pipeline defaults to VK_RASTERIZATION_ORDER_STRICT_AMD.

See Also

VkRasterizationOrderAMD, VkStructureType

Constructors

VkPipelineRasterizationStateRasterizationOrderAMD

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_RASTERIZATION_ORDER_AMD

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkRasterizationOrder :: VkRasterizationOrderAMD

rasterizationOrder must be a valid VkRasterizationOrderAMD value

newtype VkRasterizationOrderAMD

VkRasterizationOrderAMD - Specify rasterization order for a graphics pipeline

See Also

VkPipelineRasterizationStateRasterizationOrderAMD

Constructors

VkRasterizationOrderAMD Int32

pattern VK_RASTERIZATION_ORDER_STRICT_AMD :: VkRasterizationOrderAMD

VK_RASTERIZATION_ORDER_STRICT_AMD specifies that operations for each primitive in a subpass must occur in primitive order.

pattern VK_RASTERIZATION_ORDER_RELAXED_AMD :: VkRasterizationOrderAMD

VK_RASTERIZATION_ORDER_RELAXED_AMD specifies that operations for each primitive in a subpass may not occur in primitive order.

pattern VK_AMD_RASTERIZATION_ORDER_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_AMD_RASTERIZATION_ORDER_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_RASTERIZATION_ORDER_AMD :: VkStructureType