Graphics.Vulkan.C.Extensions.VK_AMD_shader_core_properties

Documentation

data VkPhysicalDeviceShaderCorePropertiesAMD

VkPhysicalDeviceShaderCorePropertiesAMD - Structure describing shader core properties that can be supported by an implementation

Members

The members of the VkPhysicalDeviceShaderCorePropertiesAMD structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceShaderCorePropertiesAMD structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

VkPhysicalDeviceShaderCorePropertiesAMD

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_CORE_PROPERTIES_AMD

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkShaderEngineCount :: Word32

shaderEngineCount is an unsigned integer value indicating the number of shader engines found inside the shader core of the physical device.

vkShaderArraysPerEngineCount :: Word32

shaderArraysPerEngineCount is an unsigned integer value indicating the number of shader arrays inside a shader engine. Each shader array has its own scan converter, set of compute units, and a render back end (color and depth buffers). Shader arrays within a shader engine share shader processor input (wave launcher) and shader export (export buffer) units. Currently, a shader engine can have one or two shader arrays.

vkComputeUnitsPerShaderArray :: Word32

computeUnitsPerShaderArray is an unsigned integer value indicating the number of compute units within a shader array. A compute unit houses a set of SIMDs along with a sequencer module and a local data store.

vkSimdPerComputeUnit :: Word32

simdPerComputeUnit is an unsigned integer value indicating the number of SIMDs inside a compute unit. Each SIMD processes a single instruction at a time.

vkWavefrontsPerSimd :: Word32

vkWavefrontSize :: Word32

wavefrontSize is an unsigned integer value indicating the number of channels (or threads) in a wavefront.

vkSgprsPerSimd :: Word32

sgprsPerSimd is an unsigned integer value indicating the number of physical Scalar General Purpose Registers (SGPRs) per SIMD.

vkMinSgprAllocation :: Word32

minSgprAllocation is an unsigned integer value indicating the minimum number of SGPRs allocated for a wave.

vkMaxSgprAllocation :: Word32

maxSgprAllocation is an unsigned integer value indicating the maximum number of SGPRs allocated for a wave.

vkSgprAllocationGranularity :: Word32

sgprAllocationGranularity is an unsigned integer value indicating the granularity of SGPR allocation for a wave.

vkVgprsPerSimd :: Word32

vgprsPerSimd is an unsigned integer value indicating the number of physical Vector General Purpose Registers (VGPRs) per SIMD.

vkMinVgprAllocation :: Word32

minVgprAllocation is an unsigned integer value indicating the minimum number of VGPRs allocated for a wave.

vkMaxVgprAllocation :: Word32

maxVgprAllocation is an unsigned integer value indicating the maximum number of VGPRs allocated for a wave.

vkVgprAllocationGranularity :: Word32

vgprAllocationGranularity is an unsigned integer value indicating the granularity of VGPR allocation for a wave.

pattern VK_AMD_SHADER_CORE_PROPERTIES_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_AMD_SHADER_CORE_PROPERTIES_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_CORE_PROPERTIES_AMD :: VkStructureType