Graphics.Vulkan.C.Extensions.VK_AMD_shader_info

Documentation

newtype VkShaderInfoTypeAMD

VkShaderInfoTypeAMD - Enum specifying which type of shader info to query

See Also

vkGetShaderInfoAMD

Constructors

VkShaderInfoTypeAMD Int32

pattern VK_SHADER_INFO_TYPE_STATISTICS_AMD :: VkShaderInfoTypeAMD

VK_SHADER_INFO_TYPE_STATISTICS_AMD specifies that device resources used by a shader will be queried.

pattern VK_SHADER_INFO_TYPE_BINARY_AMD :: VkShaderInfoTypeAMD

VK_SHADER_INFO_TYPE_BINARY_AMD specifies that implementation-specific information will be queried.

pattern VK_SHADER_INFO_TYPE_DISASSEMBLY_AMD :: VkShaderInfoTypeAMD

VK_SHADER_INFO_TYPE_DISASSEMBLY_AMD specifies that human-readable dissassembly of a shader.

data VkShaderResourceUsageAMD

VkShaderResourceUsageAMD - Resource usage information about a particular shader within a pipeline

See Also

VkShaderStatisticsInfoAMD

Constructors

VkShaderResourceUsageAMD

Fields

vkNumUsedVgprs :: Word32

numUsedVgprs is the number of vector instruction general-purpose registers used by this shader.

vkNumUsedSgprs :: Word32

numUsedSgprs is the number of scalar instruction general-purpose registers used by this shader.

vkLdsSizePerLocalWorkGroup :: Word32

ldsSizePerLocalWorkGroup is the maximum local data store size per work group in bytes.

vkLdsUsageSizeInBytes :: CSize

ldsUsageSizeInBytes is the LDS usage size in bytes per work group by this shader.

vkScratchMemUsageInBytes :: CSize

scratchMemUsageInBytes is the scratch memory usage in bytes by this shader.

data VkShaderStatisticsInfoAMD

VkShaderStatisticsInfoAMD - Statistical information about a particular shader within a pipeline

Description

Some implementations may merge multiple logical shader stages together in a single shader. In such cases, shaderStageMask will contain a bitmask of all of the stages that are active within that shader. Consequently, if specifying those stages as input to vkGetShaderInfoAMD, the same output information may be returned for all such shader stage queries.

The number of available VGPRs and SGPRs (numAvailableVgprs and numAvailableSgprs respectively) are the shader-addressable subset of physical registers that is given as a limit to the compiler for register assignment. These values may further be limited by implementations due to performance optimizations where register pressure is a bottleneck.

See Also

VkShaderResourceUsageAMD, VkShaderStageFlags

Constructors

VkShaderStatisticsInfoAMD

Fields

vkShaderStageMask :: VkShaderStageFlags

shaderStageMask are the combination of logical shader stages contained within this shader.

vkResourceUsage :: VkShaderResourceUsageAMD

resourceUsage is an instance of VkShaderResourceUsageAMD describing internal physical device resources used by this shader.

vkNumPhysicalVgprs :: Word32

numPhysicalVgprs is the maximum number of vector instruction general-purpose registers (VGPRs) available to the physical device.

vkNumPhysicalSgprs :: Word32

numPhysicalSgprs is the maximum number of scalar instruction general-purpose registers (SGPRs) available to the physical device.

vkNumAvailableVgprs :: Word32

numAvailableVgprs is the maximum limit of VGPRs made available to the shader compiler.

vkNumAvailableSgprs :: Word32

numAvailableSgprs is the maximum limit of SGPRs made available to the shader compiler.

vkComputeWorkGroupSize :: Vector 3 Word32

computeWorkGroupSize is the local workgroup size of this shader in { X, Y, Z } dimensions.

type FN_vkGetShaderInfoAMD = ("device" ::: VkDevice) -> ("pipeline" ::: VkPipeline) -> ("shaderStage" ::: VkShaderStageFlagBits) -> ("infoType" ::: VkShaderInfoTypeAMD) -> ("pInfoSize" ::: Ptr CSize) -> ("pInfo" ::: Ptr ()) -> IO VkResult

type PFN_vkGetShaderInfoAMD = FunPtr FN_vkGetShaderInfoAMD

vkGetShaderInfoAMD :: DeviceCmds -> ("device" ::: VkDevice) -> ("pipeline" ::: VkPipeline) -> ("shaderStage" ::: VkShaderStageFlagBits) -> ("infoType" ::: VkShaderInfoTypeAMD) -> ("pInfoSize" ::: Ptr CSize) -> ("pInfo" ::: Ptr ()) -> IO VkResult

vkGetShaderInfoAMD - Get information about a shader in a pipeline

Parameters

device is the device that created pipeline.

pipeline is the target of the query.

shaderStage identifies the particular shader within the pipeline about which information is being queried.

infoType describes what kind of information is being queried.

pInfoSize is a pointer to a value related to the amount of data the query returns, as described below.

pInfo is either NULL or a pointer to a buffer.

Description

If pInfo is NULL, then the maximum size of the information that can be retrieved about the shader, in bytes, is returned in pInfoSize. Otherwise, pInfoSize must point to a variable set by the user to the size of the buffer, in bytes, pointed to by pInfo, and on return the variable is overwritten with the amount of data actually written to pInfo.

If pInfoSize is less than the maximum size that can be retrieved by the pipeline cache, then at most pInfoSize bytes will be written to pInfo, and vkGetShaderInfoAMD will return VK_INCOMPLETE.

Not all information is available for every shader and implementations may not support all kinds of information for any shader. When a certain type of information is unavailable, the function returns VK_ERROR_FEATURE_NOT_PRESENT.

If information is successfully and fully queried, the function will return VK_SUCCESS.

For infoType VK_SHADER_INFO_TYPE_STATISTICS_AMD, an instance of VkShaderStatisticsInfoAMD will be written to the buffer pointed to by pInfo. This structure will be populated with statistics regarding the physical device resources used by that shader along with other miscellaneous information and is described in further detail below.

For infoType VK_SHADER_INFO_TYPE_DISASSEMBLY_AMD, pInfo points to a UTF-8 null-terminated string containing human-readable disassembly. The exact formatting and contents of the disassembly string are vendor-specific.

The formatting and contents of all other types of information, including infoType VK_SHADER_INFO_TYPE_BINARY_AMD, are left to the vendor and are not further specified by this extension.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pipeline must be a valid VkPipeline handle

shaderStage must be a valid VkShaderStageFlagBits value

infoType must be a valid VkShaderInfoTypeAMD value

pInfoSize must be a valid pointer to a size_t value

If the value referenced by pInfoSize is not 0, and pInfo is not NULL, pInfo must be a valid pointer to an array of pInfoSize bytes

pipeline must have been created, allocated, or retrieved from device

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_FEATURE_NOT_PRESENT

VK_ERROR_OUT_OF_HOST_MEMORY

See Also

VkDevice, VkPipeline, VkShaderInfoTypeAMD, VkShaderStageFlagBits

pattern VK_AMD_SHADER_INFO_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_AMD_SHADER_INFO_SPEC_VERSION :: Integral a => a