Graphics.Vulkan.C.Extensions.VK_EXT_calibrated_timestamps

Documentation

data VkCalibratedTimestampInfoEXT

VkCalibratedTimestampInfoEXT - Structure specifying the input parameters of a calibrated timestamp query

Valid Usage (Implicit)

See Also

VkStructureType, VkTimeDomainEXT, vkGetCalibratedTimestampsEXT

Constructors

VkCalibratedTimestampInfoEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_CALIBRATED_TIMESTAMP_INFO_EXT

vkPNext :: Ptr ()

pNext must be NULL

vkTimeDomain :: VkTimeDomainEXT

timeDomain must be a valid VkTimeDomainEXT value

newtype VkTimeDomainEXT

VkTimeDomainEXT - Supported time domains

See Also

VkCalibratedTimestampInfoEXT, vkGetPhysicalDeviceCalibrateableTimeDomainsEXT

Constructors

VkTimeDomainEXT Int32

pattern VK_TIME_DOMAIN_DEVICE_EXT :: VkTimeDomainEXT

VK_TIME_DOMAIN_DEVICE_EXT specifies the device time domain. Timestamp values in this time domain are comparable with device timestamp values captured using vkCmdWriteTimestamp and are defined to be incrementing according to the timestampPeriod of the device.

pattern VK_TIME_DOMAIN_CLOCK_MONOTONIC_EXT :: VkTimeDomainEXT

VK_TIME_DOMAIN_CLOCK_MONOTONIC_EXT specifies the CLOCK_MONOTONIC time domain available on POSIX platforms.

pattern VK_TIME_DOMAIN_CLOCK_MONOTONIC_RAW_EXT :: VkTimeDomainEXT

VK_TIME_DOMAIN_CLOCK_MONOTONIC_RAW_EXT specifies the CLOCK_MONOTONIC_RAW time domain available on POSIX platforms.

pattern VK_TIME_DOMAIN_QUERY_PERFORMANCE_COUNTER_EXT :: VkTimeDomainEXT

VK_TIME_DOMAIN_QUERY_PERFORMANCE_COUNTER_EXT specifies the performance counter (QPC) time domain available on Windows.

type FN_vkGetCalibratedTimestampsEXT = ("device" ::: VkDevice) -> ("timestampCount" ::: Word32) -> ("pTimestampInfos" ::: Ptr VkCalibratedTimestampInfoEXT) -> ("pTimestamps" ::: Ptr Word64) -> ("pMaxDeviation" ::: Ptr Word64) -> IO VkResult

type PFN_vkGetCalibratedTimestampsEXT = FunPtr FN_vkGetCalibratedTimestampsEXT

vkGetCalibratedTimestampsEXT :: DeviceCmds -> ("device" ::: VkDevice) -> ("timestampCount" ::: Word32) -> ("pTimestampInfos" ::: Ptr VkCalibratedTimestampInfoEXT) -> ("pTimestamps" ::: Ptr Word64) -> ("pMaxDeviation" ::: Ptr Word64) -> IO VkResult

vkGetCalibratedTimestampsEXT - Query calibrated timestamps

Parameters

device is the logical device used to perform the query.

timestampCount is the number of timestamps to query.

pTimestampInfos is a pointer to an array of timestampCount number of structures of type VkCalibratedTimestampInfoEXT, describing the time domains the calibrated timestamps should be captured from.

pTimestamps is a pointer to an array of timestampCount number of 64-bit unsigned integer values in which the requested calibrated timestamp values are returned.

pMaxDeviation is a pointer to a 64-bit unsigned integer value in which the strictly positive maximum deviation, in nanoseconds, of the calibrated timestamp values is returned.

Description

Note

The maximum deviation may vary between calls to vkGetCalibratedTimestampsEXT even for the same set of time domains due to implementation and platform specific reasons. It is the application’s responsibility to assess whether the returned maximum deviation makes the timestamp values suitable for any particular purpose and can choose to re-issue the timestamp calibration call pursuing a lower devation value.

Calibrated timestamp values can be extrapolated to estimate future coinciding timestamp values, however, depending on the nature of the time domains and other properties of the platform extrapolating values over a sufficiently long period of time may no longer be accurate enough to fit any particular purpose so applications are expected to re-calibrate the timestamps on a regular basis.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkCalibratedTimestampInfoEXT, VkDevice

type FN_vkGetPhysicalDeviceCalibrateableTimeDomainsEXT = ("physicalDevice" ::: VkPhysicalDevice) -> ("pTimeDomainCount" ::: Ptr Word32) -> ("pTimeDomains" ::: Ptr VkTimeDomainEXT) -> IO VkResult

type PFN_vkGetPhysicalDeviceCalibrateableTimeDomainsEXT = FunPtr FN_vkGetPhysicalDeviceCalibrateableTimeDomainsEXT

vkGetPhysicalDeviceCalibrateableTimeDomainsEXT :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pTimeDomainCount" ::: Ptr Word32) -> ("pTimeDomains" ::: Ptr VkTimeDomainEXT) -> IO VkResult

vkGetPhysicalDeviceCalibrateableTimeDomainsEXT - Query calibrateable time domains

Parameters

physicalDevice is the physical device from which to query the set of calibrateable time domains.

pTimeDomainCount is a pointer to an integer related to the number of calibrateable time domains available or queried, as described below.

pTimeDomains is either NULL or a pointer to an array of VkTimeDomainEXT values, indicating the supported calibrateable time domains.

Description

If pTimeDomains is NULL, then the number of calibrateable time domains supported for the given physicalDevice is returned in pTimeDomainCount. Otherwise, pTimeDomainCount must point to a variable set by the user to the number of elements in the pTimeDomains array, and on return the variable is overwritten with the number of values actually written to pTimeDomains. If the value of pTimeDomainCount is less than the number of calibrateable time domains supported, at most pTimeDomainCount values will be written to pTimeDomains. If pTimeDomainCount is smaller than the number of calibrateable time domains supported for the given physicalDevice, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pTimeDomainCount must be a valid pointer to a uint32_t value

If the value referenced by pTimeDomainCount is not 0, and pTimeDomains is not NULL, pTimeDomains must be a valid pointer to an array of pTimeDomainCount VkTimeDomainEXT values

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkPhysicalDevice, VkTimeDomainEXT

pattern VK_EXT_CALIBRATED_TIMESTAMPS_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_EXT_CALIBRATED_TIMESTAMPS_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_CALIBRATED_TIMESTAMP_INFO_EXT :: VkStructureType