Graphics.Vulkan.C.Extensions.VK_EXT_debug_marker

Documentation

data VkDebugMarkerMarkerInfoEXT

VkDebugMarkerMarkerInfoEXT - Specify parameters of a command buffer marker region

Valid Usage (Implicit)

See Also

VkStructureType, vkCmdDebugMarkerBeginEXT, vkCmdDebugMarkerInsertEXT

Constructors

VkDebugMarkerMarkerInfoEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DEBUG_MARKER_MARKER_INFO_EXT

vkPNext :: Ptr ()

pNext must be NULL

vkPMarkerName :: Ptr CChar

pMarkerName must be a null-terminated UTF-8 string

vkColor :: Vector 4 CFloat

color is an optional RGBA color value that can be associated with the marker. A particular implementation may choose to ignore this color value. The values contain RGBA values in order, in the range 0.0 to 1.0. If all elements in color are set to 0.0 then it is ignored.

data VkDebugMarkerObjectNameInfoEXT

VkDebugMarkerObjectNameInfoEXT - Specify parameters of a name to give to an object

Description

Applications may change the name associated with an object simply by calling vkDebugMarkerSetObjectNameEXT again with a new string. To remove a previously set name, pObjectName should be set to an empty string.

Valid Usage (Implicit)

See Also

VkDebugReportObjectTypeEXT, VkStructureType, vkDebugMarkerSetObjectNameEXT

Constructors

VkDebugMarkerObjectNameInfoEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DEBUG_MARKER_OBJECT_NAME_INFO_EXT

vkPNext :: Ptr ()

pNext must be NULL

vkObjectType :: VkDebugReportObjectTypeEXT

objectType must be a valid VkDebugReportObjectTypeEXT value

vkObject :: Word64

object must be a Vulkan object of the type associated with objectType as defined in https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#debug-report-object-types.

vkPObjectName :: Ptr CChar

pObjectName must be a null-terminated UTF-8 string

data VkDebugMarkerObjectTagInfoEXT

VkDebugMarkerObjectTagInfoEXT - Specify parameters of a tag to attach to an object

Description

The tagName parameter gives a name or identifier to the type of data being tagged. This can be used by debugging layers to easily filter for only data that can be used by that implementation.

Valid Usage (Implicit)

See Also

VkDebugReportObjectTypeEXT, VkStructureType, vkDebugMarkerSetObjectTagEXT

Constructors

VkDebugMarkerObjectTagInfoEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DEBUG_MARKER_OBJECT_TAG_INFO_EXT

vkPNext :: Ptr ()

pNext must be NULL

vkObjectType :: VkDebugReportObjectTypeEXT

objectType must be a valid VkDebugReportObjectTypeEXT value

vkObject :: Word64

object must be a Vulkan object of the type associated with objectType as defined in https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#debug-report-object-types.

vkTagName :: Word64

tagName is a numerical identifier of the tag.

vkTagSize :: CSize

tagSize must be greater than 0

vkPTag :: Ptr ()

pTag must be a valid pointer to an array of tagSize bytes

type FN_vkCmdDebugMarkerBeginEXT = ("commandBuffer" ::: VkCommandBuffer) -> ("pMarkerInfo" ::: Ptr VkDebugMarkerMarkerInfoEXT) -> IO ()

type PFN_vkCmdDebugMarkerBeginEXT = FunPtr FN_vkCmdDebugMarkerBeginEXT

vkCmdDebugMarkerBeginEXT :: DeviceCmds -> ("commandBuffer" ::: VkCommandBuffer) -> ("pMarkerInfo" ::: Ptr VkDebugMarkerMarkerInfoEXT) -> IO ()

vkCmdDebugMarkerBeginEXT - Open a command buffer marker region

Parameters

commandBuffer is the command buffer into which the command is recorded.

pMarkerInfo is a pointer to an instance of the VkDebugMarkerMarkerInfoEXT structure specifying the parameters of the marker region to open.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pMarkerInfo must be a valid pointer to a valid VkDebugMarkerMarkerInfoEXT structure

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkDebugMarkerMarkerInfoEXT

type FN_vkCmdDebugMarkerEndEXT = ("commandBuffer" ::: VkCommandBuffer) -> IO ()

type PFN_vkCmdDebugMarkerEndEXT = FunPtr FN_vkCmdDebugMarkerEndEXT

vkCmdDebugMarkerEndEXT :: DeviceCmds -> ("commandBuffer" ::: VkCommandBuffer) -> IO ()

vkCmdDebugMarkerEndEXT - Close a command buffer marker region

Parameters

commandBuffer is the command buffer into which the command is recorded.

Description

An application may open a marker region in one command buffer and close it in another, or otherwise split marker regions across multiple command buffers or multiple queue submissions. When viewed from the linear series of submissions to a single queue, the calls to vkCmdDebugMarkerBeginEXT and vkCmdDebugMarkerEndEXT must be matched and balanced.

Valid Usage

There must be an outstanding vkCmdDebugMarkerBeginEXT command prior to the vkCmdDebugMarkerEndEXT on the queue that commandBuffer is submitted to

If commandBuffer is a secondary command buffer, there must be an outstanding vkCmdDebugMarkerBeginEXT command recorded to commandBuffer that has not previously been ended by a call to vkCmdDebugMarkerEndEXT.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer

type FN_vkCmdDebugMarkerInsertEXT = ("commandBuffer" ::: VkCommandBuffer) -> ("pMarkerInfo" ::: Ptr VkDebugMarkerMarkerInfoEXT) -> IO ()

type PFN_vkCmdDebugMarkerInsertEXT = FunPtr FN_vkCmdDebugMarkerInsertEXT

vkCmdDebugMarkerInsertEXT :: DeviceCmds -> ("commandBuffer" ::: VkCommandBuffer) -> ("pMarkerInfo" ::: Ptr VkDebugMarkerMarkerInfoEXT) -> IO ()

vkCmdDebugMarkerInsertEXT - Insert a marker label into a command buffer

Parameters

commandBuffer is the command buffer into which the command is recorded.

pMarkerInfo is a pointer to an instance of the VkDebugMarkerMarkerInfoEXT structure specifying the parameters of the marker to insert.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pMarkerInfo must be a valid pointer to a valid VkDebugMarkerMarkerInfoEXT structure

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkDebugMarkerMarkerInfoEXT

type FN_vkDebugMarkerSetObjectNameEXT = ("device" ::: VkDevice) -> ("pNameInfo" ::: Ptr VkDebugMarkerObjectNameInfoEXT) -> IO VkResult

type PFN_vkDebugMarkerSetObjectNameEXT = FunPtr FN_vkDebugMarkerSetObjectNameEXT

vkDebugMarkerSetObjectNameEXT :: DeviceCmds -> ("device" ::: VkDevice) -> ("pNameInfo" ::: Ptr VkDebugMarkerObjectNameInfoEXT) -> IO VkResult

vkDebugMarkerSetObjectNameEXT - Give a user-friendly name to an object

Parameters

device is the device that created the object.

pNameInfo is a pointer to an instance of the VkDebugMarkerObjectNameInfoEXT structure specifying the parameters of the name to set on the object.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pNameInfo must be a valid pointer to a valid VkDebugMarkerObjectNameInfoEXT structure

Host Synchronization

Host access to pNameInfo.object must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDebugMarkerObjectNameInfoEXT, VkDevice

type FN_vkDebugMarkerSetObjectTagEXT = ("device" ::: VkDevice) -> ("pTagInfo" ::: Ptr VkDebugMarkerObjectTagInfoEXT) -> IO VkResult

type PFN_vkDebugMarkerSetObjectTagEXT = FunPtr FN_vkDebugMarkerSetObjectTagEXT

vkDebugMarkerSetObjectTagEXT :: DeviceCmds -> ("device" ::: VkDevice) -> ("pTagInfo" ::: Ptr VkDebugMarkerObjectTagInfoEXT) -> IO VkResult

vkDebugMarkerSetObjectTagEXT - Attach arbitrary data to an object

Parameters

device is the device that created the object.

pTagInfo is a pointer to an instance of the VkDebugMarkerObjectTagInfoEXT structure specifying the parameters of the tag to attach to the object.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pTagInfo must be a valid pointer to a valid VkDebugMarkerObjectTagInfoEXT structure

Host Synchronization

Host access to pTagInfo.object must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDebugMarkerObjectTagInfoEXT, VkDevice

pattern VK_EXT_DEBUG_MARKER_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_EXT_DEBUG_MARKER_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_DEBUG_MARKER_MARKER_INFO_EXT :: VkStructureType

pattern VK_STRUCTURE_TYPE_DEBUG_MARKER_OBJECT_NAME_INFO_EXT :: VkStructureType

pattern VK_STRUCTURE_TYPE_DEBUG_MARKER_OBJECT_TAG_INFO_EXT :: VkStructureType

newtype VkDebugReportObjectTypeEXT

VkDebugReportObjectTypeEXT - Specify the type of an object handle

Description

'

+------------------------------------------+---------------------------+
| 'VkDebugReportObjectTypeEXT'             | Vulkan Handle Type        |
+==========================================+===========================+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_UNKNOWN_EXT | Unknown\/Undefined Handle |
| '                                        |                           |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_INSTANCE_EX | 'Graphics.Vulkan.C.Core10 |
| T'                                       | .DeviceInitialization.VkI |
|                                          | nstance'                  |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_PHYSICAL_DE | 'Graphics.Vulkan.C.Core10 |
| VICE_EXT'                                | .DeviceInitialization.VkP |
|                                          | hysicalDevice'            |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_DEVICE_EXT' | 'Graphics.Vulkan.C.Core10 |
|                                          | .DeviceInitialization.VkD |
|                                          | evice'                    |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_QUEUE_EXT'  | 'Graphics.Vulkan.C.Core10 |
|                                          | .Queue.VkQueue'           |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_SEMAPHORE_E | 'Graphics.Vulkan.C.Core10 |
| XT'                                      | .Queue.VkSemaphore'       |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_COMMAND_BUF | 'Graphics.Vulkan.C.Core10 |
| FER_EXT'                                 | .Queue.VkCommandBuffer'   |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_FENCE_EXT'  | 'Graphics.Vulkan.C.Core10 |
|                                          | .Queue.VkFence'           |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_DEVICE_MEMO | 'Graphics.Vulkan.C.Core10 |
| RY_EXT'                                  | .Memory.VkDeviceMemory'   |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_BUFFER_EXT' | 'Graphics.Vulkan.C.Core10 |
|                                          | .MemoryManagement.VkBuffe |
|                                          | r'                        |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_IMAGE_EXT'  | 'Graphics.Vulkan.C.Core10 |
|                                          | .MemoryManagement.VkImage |
|                                          | '                         |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_EVENT_EXT'  | 'Graphics.Vulkan.C.Core10 |
|                                          | .Event.VkEvent'           |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_QUERY_POOL_ | 'Graphics.Vulkan.C.Core10 |
| EXT'                                     | .Query.VkQueryPool'       |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_BUFFER_VIEW | 'Graphics.Vulkan.C.Core10 |
| _EXT'                                    | .BufferView.VkBufferView' |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_IMAGE_VIEW_ | 'Graphics.Vulkan.C.Core10 |
| EXT'                                     | .ImageView.VkImageView'   |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_SHADER_MODU | 'Graphics.Vulkan.C.Core10 |
| LE_EXT'                                  | .Shader.VkShaderModule'   |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_PIPELINE_CA | 'Graphics.Vulkan.C.Core10 |
| CHE_EXT'                                 | .PipelineCache.VkPipeline |
|                                          | Cache'                    |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_PIPELINE_LA | 'Graphics.Vulkan.C.Core10 |
| YOUT_EXT'                                | .Pipeline.VkPipelineLayou |
|                                          | t'                        |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_RENDER_PASS | 'Graphics.Vulkan.C.Core10 |
| _EXT'                                    | .Pipeline.VkRenderPass'   |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_PIPELINE_EX | 'Graphics.Vulkan.C.Core10 |
| T'                                       | .Pipeline.VkPipeline'     |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_ | 'Graphics.Vulkan.C.Core10 |
| SET_LAYOUT_EXT'                          | .PipelineLayout.VkDescrip |
|                                          | torSetLayout'             |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_SAMPLER_EXT | 'Graphics.Vulkan.C.Core10 |
| '                                        | .Sampler.VkSampler'       |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_ | 'Graphics.Vulkan.C.Core10 |
| POOL_EXT'                                | .DescriptorSet.VkDescript |
|                                          | orPool'                   |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_ | 'Graphics.Vulkan.C.Core10 |
| SET_EXT'                                 | .DescriptorSet.VkDescript |
|                                          | orSet'                    |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_FRAMEBUFFER | 'Graphics.Vulkan.C.Core10 |
| _EXT'                                    | .Pass.VkFramebuffer'      |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_COMMAND_POO | 'Graphics.Vulkan.C.Core10 |
| L_EXT'                                   | .CommandPool.VkCommandPoo |
|                                          | l'                        |
+------------------------------------------+---------------------------+
| 'VK_DEBUG_REPORT_OBJECT_TYPE_DEBUG_REPOR | 'VkDebugReportCallbackEXT |
| T_CALLBACK_EXT_EXT'                      | '                         |
+------------------------------------------+---------------------------+

VkDebugReportObjectTypeEXT and Vulkan Handle Relationship

Note

The primary expected use of VK_ERROR_VALIDATION_FAILED_EXT is for validation layer testing. It is not expected that an application would see this error code during normal use of the validation layers.

See Also

VkDebugMarkerObjectNameInfoEXT, VkDebugMarkerObjectTagInfoEXT, vkDebugReportMessageEXT

Constructors

VkDebugReportObjectTypeEXT Int32