Graphics.Vulkan.C.Extensions.VK_EXT_depth_clip_enable

Documentation

data VkPhysicalDeviceDepthClipEnableFeaturesEXT

VkPhysicalDeviceDepthClipEnableFeaturesEXT - Structure indicating support for explicit enable of depth clip

Members

The members of the VkPhysicalDeviceDepthClipEnableFeaturesEXT structure describe the following features:

Description

If the VkPhysicalDeviceDepthClipEnableFeaturesEXT structure is included in the pNext chain of VkPhysicalDeviceFeatures2KHR, it is filled with values indicating whether the feature is supported. VkPhysicalDeviceDepthClipEnableFeaturesEXT can also be used in the pNext chain of VkDeviceCreateInfo to enable this feature.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

VkPhysicalDeviceDepthClipEnableFeaturesEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DEPTH_CLIP_ENABLE_FEATURES_EXT

vkPNext :: Ptr ()

vkDepthClipEnable :: VkBool32

depthClipEnable indicates that the implementation supports setting the depth clipping operation explicitly via the VkPipelineRasterizationDepthClipStateCreateInfoEXT pipeline state. Otherwise depth clipping is only enabled when VkPipelineRasterizationStateCreateInfo::depthClampEnable is set to VK_FALSE.

newtype VkPipelineRasterizationDepthClipStateCreateFlagsEXT

VkPipelineRasterizationDepthClipStateCreateFlagsEXT - Reserved for future use

Description

VkPipelineRasterizationDepthClipStateCreateFlagsEXT is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkPipelineRasterizationDepthClipStateCreateInfoEXT

Constructors

VkPipelineRasterizationDepthClipStateCreateFlagsEXT VkFlags

data VkPipelineRasterizationDepthClipStateCreateInfoEXT

VkPipelineRasterizationDepthClipStateCreateInfoEXT - Structure specifying depth clipping state

Valid Usage (Implicit)

See Also

VkBool32, VkPipelineRasterizationDepthClipStateCreateFlagsEXT, VkStructureType

Constructors

VkPipelineRasterizationDepthClipStateCreateInfoEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_DEPTH_CLIP_STATE_CREATE_INFO_EXT

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkFlags :: VkPipelineRasterizationDepthClipStateCreateFlagsEXT

flags must be 0

vkDepthClipEnable :: VkBool32

depthClipEnable controls whether depth clipping is enabled as described in Primitive Clipping.

pattern VK_EXT_DEPTH_CLIP_ENABLE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_EXT_DEPTH_CLIP_ENABLE_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DEPTH_CLIP_ENABLE_FEATURES_EXT :: VkStructureType

pattern VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_DEPTH_CLIP_STATE_CREATE_INFO_EXT :: VkStructureType