Graphics.Vulkan.C.Extensions.VK_EXT_display_control

Documentation

data VkDeviceEventInfoEXT

VkDeviceEventInfoEXT - Describe a device event to create

Valid Usage (Implicit)

See Also

VkDeviceEventTypeEXT, VkStructureType, vkRegisterDeviceEventEXT

Constructors

VkDeviceEventInfoEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DEVICE_EVENT_INFO_EXT

vkPNext :: Ptr ()

pNext must be NULL

vkDeviceEvent :: VkDeviceEventTypeEXT

deviceEvent must be a valid VkDeviceEventTypeEXT value

newtype VkDeviceEventTypeEXT

VkDeviceEventTypeEXT - Events that can occur on a device object

See Also

VkDeviceEventInfoEXT

Constructors

VkDeviceEventTypeEXT Int32

pattern VK_DEVICE_EVENT_TYPE_DISPLAY_HOTPLUG_EXT :: VkDeviceEventTypeEXT

VK_DEVICE_EVENT_TYPE_DISPLAY_HOTPLUG_EXT specifies that the fence is signaled when a display is plugged into or unplugged from the specified device. Applications can use this notification to determine when they need to re-enumerate the available displays on a device.

data VkDisplayEventInfoEXT

VkDisplayEventInfoEXT - Describe a display event to create

Valid Usage (Implicit)

See Also

VkDisplayEventTypeEXT, VkStructureType, vkRegisterDisplayEventEXT

Constructors

VkDisplayEventInfoEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DISPLAY_EVENT_INFO_EXT

vkPNext :: Ptr ()

pNext must be NULL

vkDisplayEvent :: VkDisplayEventTypeEXT

displayEvent must be a valid VkDisplayEventTypeEXT value

newtype VkDisplayEventTypeEXT

VkDisplayEventTypeEXT - Events that can occur on a display object

See Also

VkDisplayEventInfoEXT

Constructors

VkDisplayEventTypeEXT Int32

pattern VK_DISPLAY_EVENT_TYPE_FIRST_PIXEL_OUT_EXT :: VkDisplayEventTypeEXT

VK_DISPLAY_EVENT_TYPE_FIRST_PIXEL_OUT_EXT specifies that the fence is signaled when the first pixel of the next display refresh cycle leaves the display engine for the display.

data VkDisplayPowerInfoEXT

VkDisplayPowerInfoEXT - Describe the power state of a display

Valid Usage (Implicit)

See Also

VkDisplayPowerStateEXT, VkStructureType, vkDisplayPowerControlEXT

Constructors

VkDisplayPowerInfoEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DISPLAY_POWER_INFO_EXT

vkPNext :: Ptr ()

pNext must be NULL

vkPowerState :: VkDisplayPowerStateEXT

powerState must be a valid VkDisplayPowerStateEXT value

newtype VkDisplayPowerStateEXT

VkDisplayPowerStateEXT - Possible power states for a display

See Also

VkDisplayPowerInfoEXT

Constructors

VkDisplayPowerStateEXT Int32

pattern VK_DISPLAY_POWER_STATE_OFF_EXT :: VkDisplayPowerStateEXT

VK_DISPLAY_POWER_STATE_OFF_EXT specifies that the display is powered down.

pattern VK_DISPLAY_POWER_STATE_SUSPEND_EXT :: VkDisplayPowerStateEXT

VK_DISPLAY_POWER_STATE_SUSPEND_EXT specifies that the display is put into a low power mode, from which it may be able to transition back to VK_DISPLAY_POWER_STATE_ON_EXT more quickly than if it were in VK_DISPLAY_POWER_STATE_OFF_EXT. This state may be the same as VK_DISPLAY_POWER_STATE_OFF_EXT.

pattern VK_DISPLAY_POWER_STATE_ON_EXT :: VkDisplayPowerStateEXT

VK_DISPLAY_POWER_STATE_ON_EXT specifies that the display is powered on.

data VkSwapchainCounterCreateInfoEXT

VkSwapchainCounterCreateInfoEXT - Specify the surface counters desired

Valid Usage

The bits in surfaceCounters must be supported by VkSwapchainCreateInfoKHR::surface, as reported by vkGetPhysicalDeviceSurfaceCapabilities2EXT.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_SWAPCHAIN_COUNTER_CREATE_INFO_EXT

surfaceCounters must be a valid combination of VkSurfaceCounterFlagBitsEXT values

See Also

VkStructureType, VkSurfaceCounterFlagsEXT

Constructors

VkSwapchainCounterCreateInfoEXT

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkSurfaceCounters :: VkSurfaceCounterFlagsEXT

surfaceCounters is a bitmask of VkSurfaceCounterFlagBitsEXT specifying surface counters to enable for the swapchain.

type FN_vkDisplayPowerControlEXT = ("device" ::: VkDevice) -> ("display" ::: VkDisplayKHR) -> ("pDisplayPowerInfo" ::: Ptr VkDisplayPowerInfoEXT) -> IO VkResult

type PFN_vkDisplayPowerControlEXT = FunPtr FN_vkDisplayPowerControlEXT

vkDisplayPowerControlEXT :: DeviceCmds -> ("device" ::: VkDevice) -> ("display" ::: VkDisplayKHR) -> ("pDisplayPowerInfo" ::: Ptr VkDisplayPowerInfoEXT) -> IO VkResult

vkDisplayPowerControlEXT - Set the power state of a display

Parameters

device is a logical device associated with display.

display is the display whose power state is modified.

pDisplayPowerInfo is an instance of VkDisplayPowerInfoEXT specifying the new power state of display.

Return Codes

Success

- VK_SUCCESS

See Also

VkDevice, VkDisplayKHR, VkDisplayPowerInfoEXT

type FN_vkGetSwapchainCounterEXT = ("device" ::: VkDevice) -> ("swapchain" ::: VkSwapchainKHR) -> ("counter" ::: VkSurfaceCounterFlagBitsEXT) -> ("pCounterValue" ::: Ptr Word64) -> IO VkResult

type PFN_vkGetSwapchainCounterEXT = FunPtr FN_vkGetSwapchainCounterEXT

vkGetSwapchainCounterEXT :: DeviceCmds -> ("device" ::: VkDevice) -> ("swapchain" ::: VkSwapchainKHR) -> ("counter" ::: VkSurfaceCounterFlagBitsEXT) -> ("pCounterValue" ::: Ptr Word64) -> IO VkResult

vkGetSwapchainCounterEXT - Query the current value of a surface counter

Parameters

device is the VkDevice associated with swapchain.

swapchain is the swapchain from which to query the counter value.

counter is the counter to query.

pCounterValue will return the current value of the counter.

Description

If a counter is not available because the swapchain is out of date, the implementation may return VK_ERROR_OUT_OF_DATE_KHR.

Valid Usage

One or more present commands on swapchain must have been processed by the presentation engine.

Valid Usage (Implicit)

device must be a valid VkDevice handle

swapchain must be a valid VkSwapchainKHR handle

counter must be a valid VkSurfaceCounterFlagBitsEXT value

pCounterValue must be a valid pointer to a uint64_t value

Both of device, and swapchain must have been created, allocated, or retrieved from the same VkInstance

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_DEVICE_LOST

VK_ERROR_OUT_OF_DATE_KHR

See Also

VkDevice, VkSurfaceCounterFlagBitsEXT, VkSwapchainKHR

type FN_vkRegisterDeviceEventEXT = ("device" ::: VkDevice) -> ("pDeviceEventInfo" ::: Ptr VkDeviceEventInfoEXT) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pFence" ::: Ptr VkFence) -> IO VkResult

type PFN_vkRegisterDeviceEventEXT = FunPtr FN_vkRegisterDeviceEventEXT

vkRegisterDeviceEventEXT :: DeviceCmds -> ("device" ::: VkDevice) -> ("pDeviceEventInfo" ::: Ptr VkDeviceEventInfoEXT) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pFence" ::: Ptr VkFence) -> IO VkResult

vkRegisterDeviceEventEXT - Signal a fence when a device event occurs

Parameters

device is a logical device on which the event may occur.

pDeviceEventInfo is a pointer to an instance of the VkDeviceEventInfoEXT structure describing the event of interest to the application.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pFence points to a handle in which the resulting fence object is returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pDeviceEventInfo must be a valid pointer to a valid VkDeviceEventInfoEXT structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pFence must be a valid pointer to a VkFence handle

Return Codes

Success

- VK_SUCCESS

See Also

VkAllocationCallbacks, VkDevice, VkDeviceEventInfoEXT, VkFence

type FN_vkRegisterDisplayEventEXT = ("device" ::: VkDevice) -> ("display" ::: VkDisplayKHR) -> ("pDisplayEventInfo" ::: Ptr VkDisplayEventInfoEXT) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pFence" ::: Ptr VkFence) -> IO VkResult

type PFN_vkRegisterDisplayEventEXT = FunPtr FN_vkRegisterDisplayEventEXT

vkRegisterDisplayEventEXT :: DeviceCmds -> ("device" ::: VkDevice) -> ("display" ::: VkDisplayKHR) -> ("pDisplayEventInfo" ::: Ptr VkDisplayEventInfoEXT) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pFence" ::: Ptr VkFence) -> IO VkResult

vkRegisterDisplayEventEXT - Signal a fence when a display event occurs

Parameters

device is a logical device associated with display

display is the display on which the event may occur.

pDisplayEventInfo is a pointer to an instance of the VkDisplayEventInfoEXT structure describing the event of interest to the application.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pFence points to a handle in which the resulting fence object is returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

display must be a valid VkDisplayKHR handle

pDisplayEventInfo must be a valid pointer to a valid VkDisplayEventInfoEXT structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pFence must be a valid pointer to a VkFence handle

Return Codes

Success

- VK_SUCCESS

See Also

VkAllocationCallbacks, VkDevice, VkDisplayEventInfoEXT, VkDisplayKHR, VkFence

pattern VK_EXT_DISPLAY_CONTROL_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_EXT_DISPLAY_CONTROL_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_DEVICE_EVENT_INFO_EXT :: VkStructureType

pattern VK_STRUCTURE_TYPE_DISPLAY_EVENT_INFO_EXT :: VkStructureType

pattern VK_STRUCTURE_TYPE_DISPLAY_POWER_INFO_EXT :: VkStructureType

pattern VK_STRUCTURE_TYPE_SWAPCHAIN_COUNTER_CREATE_INFO_EXT :: VkStructureType