Graphics.Vulkan.C.Extensions.VK_EXT_display_surface_counter

Documentation

data VkSurfaceCapabilities2EXT

VkSurfaceCapabilities2EXT - Structure describing capabilities of a surface

Members

All members of VkSurfaceCapabilities2EXT are identical to the corresponding members of VkSurfaceCapabilitiesKHR where one exists. The remaining members are:

Valid Usage (Implicit)

See Also

VkCompositeAlphaFlagsKHR, VkExtent2D, VkImageUsageFlags, VkStructureType, VkSurfaceCounterFlagsEXT, VkSurfaceTransformFlagBitsKHR, VkSurfaceTransformFlagsKHR, vkGetPhysicalDeviceSurfaceCapabilities2EXT

Constructors

VkSurfaceCapabilities2EXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_SURFACE_CAPABILITIES_2_EXT

vkPNext :: Ptr ()

pNext must be NULL

vkMinImageCount :: Word32

vkMaxImageCount :: Word32

vkCurrentExtent :: VkExtent2D

vkMinImageExtent :: VkExtent2D

vkMaxImageExtent :: VkExtent2D

vkMaxImageArrayLayers :: Word32

vkSupportedTransforms :: VkSurfaceTransformFlagsKHR

vkCurrentTransform :: VkSurfaceTransformFlagBitsKHR

vkSupportedCompositeAlpha :: VkCompositeAlphaFlagsKHR

vkSupportedUsageFlags :: VkImageUsageFlags

vkSupportedSurfaceCounters :: VkSurfaceCounterFlagsEXT

supportedSurfaceCounters must not include VK_SURFACE_COUNTER_VBLANK_EXT unless the surface queried is a display surface.

newtype VkSurfaceCounterFlagBitsEXT

VkSurfaceCounterFlagBitsEXT - Surface-relative counter types

See Also

VkSurfaceCounterFlagsEXT, vkGetSwapchainCounterEXT

Constructors

VkSurfaceCounterFlagBitsEXT VkFlags

pattern VK_SURFACE_COUNTER_VBLANK_EXT :: VkSurfaceCounterFlagBitsEXT

VK_SURFACE_COUNTER_VBLANK_EXT specifies a counter incrementing once every time a vertical blanking period occurs on the display associated with the surface.

type VkSurfaceCounterFlagsEXT = VkSurfaceCounterFlagBitsEXT

VkSurfaceCounterFlagsEXT - Bitmask of VkSurfaceCounterFlagBitsEXT

Description

VkSurfaceCounterFlagsEXT is a bitmask type for setting a mask of zero or more VkSurfaceCounterFlagBitsEXT.

See Also

VkSurfaceCapabilities2EXT, VkSurfaceCounterFlagBitsEXT, VkSwapchainCounterCreateInfoEXT

type FN_vkGetPhysicalDeviceSurfaceCapabilities2EXT = ("physicalDevice" ::: VkPhysicalDevice) -> ("surface" ::: VkSurfaceKHR) -> ("pSurfaceCapabilities" ::: Ptr VkSurfaceCapabilities2EXT) -> IO VkResult

type PFN_vkGetPhysicalDeviceSurfaceCapabilities2EXT = FunPtr FN_vkGetPhysicalDeviceSurfaceCapabilities2EXT

vkGetPhysicalDeviceSurfaceCapabilities2EXT :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("surface" ::: VkSurfaceKHR) -> ("pSurfaceCapabilities" ::: Ptr VkSurfaceCapabilities2EXT) -> IO VkResult

vkGetPhysicalDeviceSurfaceCapabilities2EXT - Query surface capabilities

Parameters

physicalDevice is the physical device that will be associated with the swapchain to be created, as described for vkCreateSwapchainKHR.

surface is the surface that will be associated with the swapchain.

pSurfaceCapabilities is a pointer to an instance of the VkSurfaceCapabilities2EXT structure in which the capabilities are returned.

Description

vkGetPhysicalDeviceSurfaceCapabilities2EXT behaves similarly to vkGetPhysicalDeviceSurfaceCapabilitiesKHR, with the ability to return extended information by adding extension structures to the pNext chain of its pSurfaceCapabilities parameter.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

surface must be a valid VkSurfaceKHR handle

pSurfaceCapabilities must be a valid pointer to a VkSurfaceCapabilities2EXT structure

Both of physicalDevice, and surface must have been created, allocated, or retrieved from the same VkInstance

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

See Also

VkPhysicalDevice, VkSurfaceCapabilities2EXT, VkSurfaceKHR

pattern VK_EXT_DISPLAY_SURFACE_COUNTER_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_EXT_DISPLAY_SURFACE_COUNTER_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_SURFACE_CAPABILITIES2_EXT :: VkStructureType

pattern VK_STRUCTURE_TYPE_SURFACE_CAPABILITIES_2_EXT :: VkStructureType