Graphics.Vulkan.C.Extensions.VK_EXT_hdr_metadata

Documentation

data VkHdrMetadataEXT

VkHdrMetadataEXT - structure to specify Hdr metadata

Valid Usage (Implicit)

Note

The validity and use of this data is outside the scope of Vulkan.

See Also

VkStructureType, VkXYColorEXT, vkSetHdrMetadataEXT

Constructors

VkHdrMetadataEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_HDR_METADATA_EXT

vkPNext :: Ptr ()

pNext must be NULL

vkDisplayPrimaryRed :: VkXYColorEXT

displayPrimaryRed is the mastering display’s red primary in chromaticity coordinates

vkDisplayPrimaryGreen :: VkXYColorEXT

displayPrimaryGreen is the mastering display’s green primary in chromaticity coordinates

vkDisplayPrimaryBlue :: VkXYColorEXT

displayPrimaryBlue is the mastering display’s blue primary in chromaticity coordinates

vkWhitePoint :: VkXYColorEXT

whitePoint is the mastering display’s white-point in chromaticity coordinates

vkMaxLuminance :: CFloat

maxLuminance is the maximum luminance of the mastering display in nits

vkMinLuminance :: CFloat

minLuminance is the minimum luminance of the mastering display in nits

vkMaxContentLightLevel :: CFloat

maxContentLightLevel is content’s maximum luminance in nits

vkMaxFrameAverageLightLevel :: CFloat

maxFrameAverageLightLevel is the maximum frame average light level in nits

data VkXYColorEXT

VkXYColorEXT - structure to specify X,Y chromaticity coordinates

See Also

VkHdrMetadataEXT

Constructors

VkXYColorEXT

Fields

vkX :: CFloat

vkY :: CFloat

type FN_vkSetHdrMetadataEXT = ("device" ::: VkDevice) -> ("swapchainCount" ::: Word32) -> ("pSwapchains" ::: Ptr VkSwapchainKHR) -> ("pMetadata" ::: Ptr VkHdrMetadataEXT) -> IO ()

type PFN_vkSetHdrMetadataEXT = FunPtr FN_vkSetHdrMetadataEXT

vkSetHdrMetadataEXT :: DeviceCmds -> ("device" ::: VkDevice) -> ("swapchainCount" ::: Word32) -> ("pSwapchains" ::: Ptr VkSwapchainKHR) -> ("pMetadata" ::: Ptr VkHdrMetadataEXT) -> IO ()

vkSetHdrMetadataEXT - function to set Hdr metadata

Parameters

device is the logical device where the swapchain(s) were created.

swapchainCount is the number of swapchains included in pSwapchains.

pSwapchains is a pointer to the array of swapchainCount VkSwapchainKHR handles.

pMetadata is a pointer to the array of swapchainCount VkHdrMetadataEXT structures.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pSwapchains must be a valid pointer to an array of swapchainCount valid VkSwapchainKHR handles

pMetadata must be a valid pointer to an array of swapchainCount valid VkHdrMetadataEXT structures

swapchainCount must be greater than 0

Both of device, and the elements of pSwapchains must have been created, allocated, or retrieved from the same VkInstance

See Also

VkDevice, VkHdrMetadataEXT, VkSwapchainKHR

pattern VK_EXT_HDR_METADATA_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_EXT_HDR_METADATA_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_HDR_METADATA_EXT :: VkStructureType