Graphics.Vulkan.C.Extensions.VK_EXT_inline_uniform_block

Documentation

data VkDescriptorPoolInlineUniformBlockCreateInfoEXT

VkDescriptorPoolInlineUniformBlockCreateInfoEXT - Structure specifying the maximum number of inline uniform block bindings of a newly created descriptor pool

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

VkDescriptorPoolInlineUniformBlockCreateInfoEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_INLINE_UNIFORM_BLOCK_CREATE_INFO_EXT

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkMaxInlineUniformBlockBindings :: Word32

maxInlineUniformBlockBindings is the number of inline uniform block bindings to allocate.

data VkPhysicalDeviceInlineUniformBlockFeaturesEXT

VkPhysicalDeviceInlineUniformBlockFeaturesEXT - Structure describing inline uniform block features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceInlineUniformBlockFeaturesEXT structure describe the following features:

Description

If the VkPhysicalDeviceInlineUniformBlockFeaturesEXT structure is included in the pNext chain of VkPhysicalDeviceFeatures2, it is filled with values indicating whether each feature is supported. VkPhysicalDeviceInlineUniformBlockFeaturesEXT can also be used in the pNext chain of VkDeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

VkPhysicalDeviceInlineUniformBlockFeaturesEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_INLINE_UNIFORM_BLOCK_FEATURES_EXT

vkPNext :: Ptr ()

vkInlineUniformBlock :: VkBool32

inlineUniformBlock indicates whether the implementation supports inline uniform block descriptors. If this feature is not enabled, VK_DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT must not be used.

vkDescriptorBindingInlineUniformBlockUpdateAfterBind :: VkBool32

descriptorBindingInlineUniformBlockUpdateAfterBind is VK_FALSE and reserved for future use.

data VkPhysicalDeviceInlineUniformBlockPropertiesEXT

VkPhysicalDeviceInlineUniformBlockPropertiesEXT - Structure describing inline uniform block properties that can be supported by an implementation

Members

The members of the VkPhysicalDeviceInlineUniformBlockPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceInlineUniformBlockPropertiesEXT structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

VkPhysicalDeviceInlineUniformBlockPropertiesEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_INLINE_UNIFORM_BLOCK_PROPERTIES_EXT

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkMaxInlineUniformBlockSize :: Word32

maxInlineUniformBlockSize is the maximum size in bytes of an inline uniform block binding.

vkMaxPerStageDescriptorInlineUniformBlocks :: Word32

vkMaxPerStageDescriptorUpdateAfterBindInlineUniformBlocks :: Word32

maxPerStageDescriptorUpdateAfterBindInlineUniformBlocks equals maxPerStageDescriptorInlineUniformBlocks and is reserved for future use.

vkMaxDescriptorSetInlineUniformBlocks :: Word32

maxDescriptorSetInlineUniformBlocks is the maximum number of inline uniform block bindings that can be included in descriptor bindings in a pipeline layout across all pipeline shader stages and descriptor set numbers. Descriptor bindings with a descriptor type of VK_DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT count against this limit.

vkMaxDescriptorSetUpdateAfterBindInlineUniformBlocks :: Word32

maxDescriptorSetUpdateAfterBindInlineUniformBlocks equals maxDescriptorSetInlineUniformBlocks and is reserved for future use.

data VkWriteDescriptorSetInlineUniformBlockEXT

VkWriteDescriptorSetInlineUniformBlockEXT - Structure specifying inline uniform block data

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

VkWriteDescriptorSetInlineUniformBlockEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET_INLINE_UNIFORM_BLOCK_EXT

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkDataSize :: Word32

dataSize must be greater than 0

vkPData :: Ptr ()

pData must be a valid pointer to an array of dataSize bytes

pattern VK_DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT :: VkDescriptorType

pattern VK_EXT_INLINE_UNIFORM_BLOCK_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_EXT_INLINE_UNIFORM_BLOCK_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_INLINE_UNIFORM_BLOCK_CREATE_INFO_EXT :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_INLINE_UNIFORM_BLOCK_FEATURES_EXT :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_INLINE_UNIFORM_BLOCK_PROPERTIES_EXT :: VkStructureType

pattern VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET_INLINE_UNIFORM_BLOCK_EXT :: VkStructureType