Graphics.Vulkan.C.Extensions.VK_EXT_memory_priority

Documentation

data VkMemoryPriorityAllocateInfoEXT

VkMemoryPriorityAllocateInfoEXT - Specify a memory allocation priority

Description

Memory allocations with higher priority may be more likely to stay in device-local memory when the system is under memory pressure.

If this structure is not included, it is as if the priority value were 0.5.

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

VkMemoryPriorityAllocateInfoEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_MEMORY_PRIORITY_ALLOCATE_INFO_EXT

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkPriority :: CFloat

priority must be between 0 and 1, inclusive

data VkPhysicalDeviceMemoryPriorityFeaturesEXT

VkPhysicalDeviceMemoryPriorityFeaturesEXT - Structure describing memory priority features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceMemoryPriorityFeaturesEXT structure describe the following features:

Description

If the VkPhysicalDeviceMemoryPriorityFeaturesEXT structure is included in the pNext chain of VkPhysicalDeviceFeatures2KHR, it is filled with values indicating whether the feature is supported. VkPhysicalDeviceMemoryPriorityFeaturesEXT can also be used in the pNext chain of VkDeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

VkPhysicalDeviceMemoryPriorityFeaturesEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MEMORY_PRIORITY_FEATURES_EXT

vkPNext :: Ptr ()

vkMemoryPriority :: VkBool32

memoryPriority indicates that the implementation supports memory priorities specified at memory allocation time via VkMemoryPriorityAllocateInfoEXT.

pattern VK_EXT_MEMORY_PRIORITY_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_EXT_MEMORY_PRIORITY_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_MEMORY_PRIORITY_ALLOCATE_INFO_EXT :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MEMORY_PRIORITY_FEATURES_EXT :: VkStructureType