Graphics.Vulkan.C.Extensions.VK_EXT_pipeline_creation_feedback

Documentation

data VkPipelineCreationFeedbackCreateInfoEXT

VkPipelineCreationFeedbackCreateInfoEXT - Request for feedback about the creation of a pipeline

Description

An implementation should write pipeline creation feedback to pPipelineCreationFeedback and may write pipeline stage creation feedback to pPipelineStageCreationFeedbacks. An implementation must set or clear the VK_PIPELINE_CREATION_FEEDBACK_VALID_BIT_EXT in VkPipelineCreationFeedbackEXT::flags for pPipelineCreationFeedback and every element of pPipelineStageCreationFeedbacks.

Note

One common scenario for an implementation to skip per-stage feedback is when VK_PIPELINE_CREATION_FEEDBACK_APPLICATION_PIPELINE_CACHE_HIT_BIT_EXT is set in pPipelineCreationFeedback.

When chained to VkGraphicsPipelineCreateInfo, the i element of pPipelineStageCreationFeedbacks corresponds to the i element of VkGraphicsPipelineCreateInfo::pStages. When chained to VkComputePipelineCreateInfo, the first element of pPipelineStageCreationFeedbacks corresponds to VkComputePipelineCreateInfo::stage.

Valid Usage

When chained to VkGraphicsPipelineCreateInfo, VkPipelineCreationFeedbackEXT::pipelineStageCreationFeedbackCount must equal VkGraphicsPipelineCreateInfo::stageCount

When chained to VkComputePipelineCreateInfo, VkPipelineCreationFeedbackEXT::pipelineStageCreationFeedbackCount must equal 1

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PIPELINE_CREATION_FEEDBACK_CREATE_INFO_EXT

pPipelineCreationFeedback must be a valid pointer to a VkPipelineCreationFeedbackEXT structure

pPipelineStageCreationFeedbacks must be a valid pointer to an array of pipelineStageCreationFeedbackCount VkPipelineCreationFeedbackEXT structures

pipelineStageCreationFeedbackCount must be greater than 0

See Also

VkComputePipelineCreateInfo, VkGraphicsPipelineCreateInfo, VkPipelineCreationFeedbackEXT, VkRayTracingPipelineCreateInfoNV, VkStructureType

Constructors

VkPipelineCreationFeedbackCreateInfoEXT

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkPPipelineCreationFeedback :: Ptr VkPipelineCreationFeedbackEXT

pPipelineCreationFeedback is a pointer to a VkPipelineCreationFeedbackEXT structure.

vkPipelineStageCreationFeedbackCount :: Word32

pipelineStageCreationFeedbackCount is the number of elements in pPipelineStageCreationFeedbacks.

vkPPipelineStageCreationFeedbacks :: Ptr VkPipelineCreationFeedbackEXT

pPipelineStageCreationFeedbacks is an array of size pipelineStageCreationFeedbackCount of VkPipelineCreationFeedbackEXT structures.

data VkPipelineCreationFeedbackEXT

VkPipelineCreationFeedbackEXT - Feedback about the creation of a pipeline or pipeline stage

Description

If the VK_PIPELINE_CREATION_FEEDBACK_VALID_BIT_EXT is not set in flags, an implementation must not set any other bits in flags, and all other VkPipelineCreationFeedbackEXT data members are undefined.

See Also

VkPipelineCreationFeedbackCreateInfoEXT, VkPipelineCreationFeedbackFlagBitsEXT, VkPipelineCreationFeedbackFlagsEXT

Constructors

VkPipelineCreationFeedbackEXT

Fields

vkFlags :: VkPipelineCreationFeedbackFlagsEXT

flags is a bitmask of VkPipelineCreationFeedbackFlagBitsEXT providing feedback about the creation of a pipeline or of a pipeline stage.

vkDuration :: Word64

duration is the duration spent creating a pipeline or pipeline stage in nanoseconds.

newtype VkPipelineCreationFeedbackFlagBitsEXT

VkPipelineCreationFeedbackFlagBitsEXT - Bitmask specifying pipeline or pipeline stage creation feedback

See Also

VkPipelineCreationFeedbackCreateInfoEXT, VkPipelineCreationFeedbackEXT, VkPipelineCreationFeedbackFlagsEXT

Constructors

VkPipelineCreationFeedbackFlagBitsEXT VkFlags

pattern VK_PIPELINE_CREATION_FEEDBACK_VALID_BIT_EXT :: VkPipelineCreationFeedbackFlagBitsEXT

VK_PIPELINE_CREATION_FEEDBACK_VALID_BIT_EXT indicates that the feedback information is valid.

pattern VK_PIPELINE_CREATION_FEEDBACK_APPLICATION_PIPELINE_CACHE_HIT_BIT_EXT :: VkPipelineCreationFeedbackFlagBitsEXT

VK_PIPELINE_CREATION_FEEDBACK_APPLICATION_PIPELINE_CACHE_HIT_BIT_EXT indicates that a readily usable pipeline or pipeline stage was found in the pipelineCache specified by the application in the pipeline creation command.

An implementation should set the VK_PIPELINE_CREATION_FEEDBACK_APPLICATION_PIPELINE_CACHE_HIT_BIT_EXT bit if it was able to avoid the large majority of pipeline or pipeline stage creation work by using the pipelineCache parameter of vkCreateGraphicsPipelines, or vkCreateComputePipelines. When an implementation sets this bit for the entire pipeline, it may leave it unset for any stage.

Note

Implementations are encouraged to provide a meaningful signal to applications using this bit. The intention is to communicate to the application that the pipeline or pipeline stage was created "as fast as it gets" using the pipeline cache provided by the application. If an implementation uses an internal cache, it is discouraged from setting this bit as the feedback would be unactionable.

pattern VK_PIPELINE_CREATION_FEEDBACK_BASE_PIPELINE_ACCELERATION_BIT_EXT :: VkPipelineCreationFeedbackFlagBitsEXT

VK_PIPELINE_CREATION_FEEDBACK_BASE_PIPELINE_ACCELERATION_BIT_EXT indicates that the base pipeline specified by the basePipelineHandle or basePipelineIndex member of the Vk*PipelineCreateInfo structure was used to accelerate the creation of the pipeline.

An implementation should set the VK_PIPELINE_CREATION_FEEDBACK_BASE_PIPELINE_ACCELERATION_BIT_EXT bit if it was able to avoid a significant amount of work by using the base pipeline.

Note

While "significant amount of work" is subjective, implementations are encouraged to provide a meaningful signal to applications using this bit. For example, a 1% reduction in duration may not warrant setting this bit, while a 50% reduction would.

type VkPipelineCreationFeedbackFlagsEXT = VkPipelineCreationFeedbackFlagBitsEXT

VkPipelineCreationFeedbackFlagsEXT - Bitmask of VkPipelineCreationFeedbackFlagBitsEXT

Description

VkPipelineCreationFeedbackFlagsEXT is a bitmask type for providing zero or more VkPipelineCreationFeedbackFlagBitsEXT.

See Also

VkPipelineCreationFeedbackEXT, VkPipelineCreationFeedbackFlagBitsEXT

pattern VK_EXT_PIPELINE_CREATION_FEEDBACK_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_EXT_PIPELINE_CREATION_FEEDBACK_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_PIPELINE_CREATION_FEEDBACK_CREATE_INFO_EXT :: VkStructureType