Graphics.Vulkan.C.Extensions.VK_EXT_validation_features

Documentation

newtype VkValidationFeatureDisableEXT

VkValidationFeatureDisableEXT - Specify validation features to disable

See Also

VkValidationFeaturesEXT

Constructors

VkValidationFeatureDisableEXT Int32

pattern VK_VALIDATION_FEATURE_DISABLE_ALL_EXT :: VkValidationFeatureDisableEXT

VK_VALIDATION_FEATURE_DISABLE_ALL_EXT specifies that all validation checks are disabled.

pattern VK_VALIDATION_FEATURE_DISABLE_SHADERS_EXT :: VkValidationFeatureDisableEXT

VK_VALIDATION_FEATURE_DISABLE_SHADERS_EXT specifies that shader validation is disabled. This feature is enabled by default.

pattern VK_VALIDATION_FEATURE_DISABLE_THREAD_SAFETY_EXT :: VkValidationFeatureDisableEXT

VK_VALIDATION_FEATURE_DISABLE_THREAD_SAFETY_EXT specifies that thread safety validation is disabled. This feature is enabled by default.

pattern VK_VALIDATION_FEATURE_DISABLE_API_PARAMETERS_EXT :: VkValidationFeatureDisableEXT

VK_VALIDATION_FEATURE_DISABLE_API_PARAMETERS_EXT specifies that stateless parameter validation is disabled. This feature is enabled by default.

pattern VK_VALIDATION_FEATURE_DISABLE_OBJECT_LIFETIMES_EXT :: VkValidationFeatureDisableEXT

VK_VALIDATION_FEATURE_DISABLE_OBJECT_LIFETIMES_EXT specifies that object lifetime validation is disabled. This feature is enabled by default.

pattern VK_VALIDATION_FEATURE_DISABLE_CORE_CHECKS_EXT :: VkValidationFeatureDisableEXT

VK_VALIDATION_FEATURE_DISABLE_CORE_CHECKS_EXT specifies that core validation checks are disabled. This feature is enabled by default. If this feature is disabled, the shader validation and GPU-assisted validation features are also disabled.

pattern VK_VALIDATION_FEATURE_DISABLE_UNIQUE_HANDLES_EXT :: VkValidationFeatureDisableEXT

VK_VALIDATION_FEATURE_DISABLE_UNIQUE_HANDLES_EXT specifies that protection against duplicate non-dispatchable object handles is disabled. This feature is enabled by default.

newtype VkValidationFeatureEnableEXT

VkValidationFeatureEnableEXT - Specify validation features to enable

See Also

VkValidationFeaturesEXT

Constructors

VkValidationFeatureEnableEXT Int32

pattern VK_VALIDATION_FEATURE_ENABLE_GPU_ASSISTED_EXT :: VkValidationFeatureEnableEXT

VK_VALIDATION_FEATURE_ENABLE_GPU_ASSISTED_EXT specifies that GPU-assisted validation is enabled. Activating this feature instruments shader programs to generate additional diagnostic data. This feature is disabled by default.

pattern VK_VALIDATION_FEATURE_ENABLE_GPU_ASSISTED_RESERVE_BINDING_SLOT_EXT :: VkValidationFeatureEnableEXT

VK_VALIDATION_FEATURE_ENABLE_GPU_ASSISTED_RESERVE_BINDING_SLOT_EXT specifies that the validation layers reserve a descriptor set binding slot for their own use. The layer reports a value for VkPhysicalDeviceLimits::maxBoundDescriptorSets that is one less than the value reported by the device. If the device supports the binding of only one descriptor set, the validation layer does not perform GPU-assisted validation. This feature is disabled by default. The GPU-assisted validation feature must be enabled in order to use this feature.

data VkValidationFeaturesEXT

VkValidationFeaturesEXT - Specify validation features to enable or disable for a Vulkan instance

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_VALIDATION_FEATURES_EXT

If enabledValidationFeatureCount is not 0, pEnabledValidationFeatures must be a valid pointer to an array of enabledValidationFeatureCount valid VkValidationFeatureEnableEXT values

If disabledValidationFeatureCount is not 0, pDisabledValidationFeatures must be a valid pointer to an array of disabledValidationFeatureCount valid VkValidationFeatureDisableEXT values

See Also

VkStructureType, VkValidationFeatureDisableEXT, VkValidationFeatureEnableEXT

Constructors

VkValidationFeaturesEXT

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkEnabledValidationFeatureCount :: Word32

enabledValidationFeatureCount is the number of features to enable.

vkPEnabledValidationFeatures :: Ptr VkValidationFeatureEnableEXT

pEnabledValidationFeatures is a pointer to an array of VkValidationFeatureEnableEXT values specifying the validation features to be enabled.

vkDisabledValidationFeatureCount :: Word32

disabledValidationFeatureCount is the number of features to disable.

vkPDisabledValidationFeatures :: Ptr VkValidationFeatureDisableEXT

pDisabledValidationFeatures is a pointer to an array of VkValidationFeatureDisableEXT values specifying the validation features to be disabled.

pattern VK_EXT_VALIDATION_FEATURES_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_EXT_VALIDATION_FEATURES_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_VALIDATION_FEATURES_EXT :: VkStructureType