Graphics.Vulkan.C.Extensions.VK_EXT_validation_flags

Documentation

newtype VkValidationCheckEXT

VkValidationCheckEXT - Specify validation checks to disable

See Also

VkValidationFlagsEXT

Constructors

VkValidationCheckEXT Int32

pattern VK_VALIDATION_CHECK_ALL_EXT :: VkValidationCheckEXT

VK_VALIDATION_CHECK_ALL_EXT specifies that all validation checks are disabled.

pattern VK_VALIDATION_CHECK_SHADERS_EXT :: VkValidationCheckEXT

VK_VALIDATION_CHECK_SHADERS_EXT specifies that shader validation is disabled.

data VkValidationFlagsEXT

VkValidationFlagsEXT - Specify validation checks to disable for a Vulkan instance

Valid Usage (Implicit)

See Also

VkStructureType, VkValidationCheckEXT

Constructors

VkValidationFlagsEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_VALIDATION_FLAGS_EXT

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkDisabledValidationCheckCount :: Word32

disabledValidationCheckCount must be greater than 0

vkPDisabledValidationChecks :: Ptr VkValidationCheckEXT

pDisabledValidationChecks must be a valid pointer to an array of disabledValidationCheckCount valid VkValidationCheckEXT values

pattern VK_EXT_VALIDATION_FLAGS_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_EXT_VALIDATION_FLAGS_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_VALIDATION_FLAGS_EXT :: VkStructureType