Graphics.Vulkan.C.Extensions.VK_EXT_vertex_attribute_divisor

Documentation

data VkPhysicalDeviceVertexAttributeDivisorFeaturesEXT

VkPhysicalDeviceVertexAttributeDivisorFeaturesEXT - Structure describing if fetching of vertex attribute may be repeated for instanced rendering

Description

If the VkPhysicalDeviceVertexAttributeDivisorFeaturesEXT structure is included in the pNext chain of VkPhysicalDeviceFeatures2, it is filled with values indicating the implementation-dependent behavior. VkPhysicalDeviceVertexAttributeDivisorFeaturesEXT can also be used in pNext chain of VkDeviceCreateInfo to enable the feature.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

VkPhysicalDeviceVertexAttributeDivisorFeaturesEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VERTEX_ATTRIBUTE_DIVISOR_FEATURES_EXT

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkVertexAttributeInstanceRateDivisor :: VkBool32

vertexAttributeInstanceRateDivisor specifies whether vertex attribute fetching may be repeated in case of instanced rendering.

vkVertexAttributeInstanceRateZeroDivisor :: VkBool32

vertexAttributeInstanceRateZeroDivisor specifies whether a zero value for VkVertexInputBindingDivisorDescriptionEXT::divisor is supported.

data VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT

VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT - Structure describing max value of vertex attribute divisor that can be supported by an implementation

Members

The members of the VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VERTEX_ATTRIBUTE_DIVISOR_PROPERTIES_EXT

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkMaxVertexAttribDivisor :: Word32

maxVertexAttribDivisor is the maximum value of the number of instances that will repeat the value of vertex attribute data when instanced rendering is enabled.

data VkPipelineVertexInputDivisorStateCreateInfoEXT

VkPipelineVertexInputDivisorStateCreateInfoEXT - Structure specifying vertex attributes assignment during instanced rendering

Valid Usage (Implicit)

See Also

VkStructureType, VkVertexInputBindingDivisorDescriptionEXT

Constructors

VkPipelineVertexInputDivisorStateCreateInfoEXT

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_DIVISOR_STATE_CREATE_INFO_EXT

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure

vkVertexBindingDivisorCount :: Word32

vertexBindingDivisorCount must be greater than 0

vkPVertexBindingDivisors :: Ptr VkVertexInputBindingDivisorDescriptionEXT

pVertexBindingDivisors must be a valid pointer to an array of vertexBindingDivisorCount VkVertexInputBindingDivisorDescriptionEXT structures

data VkVertexInputBindingDivisorDescriptionEXT

VkVertexInputBindingDivisorDescriptionEXT - Structure specifying a divisor used in instanced rendering

Description

If this structure is not used to define a divisor value for an attribute then the divisor has a logical default value of 1.

Valid Usage

binding must be less than VkPhysicalDeviceLimits::maxVertexInputBindings

If the vertexAttributeInstanceRateZeroDivisor feature is not enabled, divisor must not be 0

If the vertexAttributeInstanceRateDivisor feature is not enabled, divisor must be 1

divisor must be a value between 0 and VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT::maxVertexAttribDivisor, inclusive.

VkVertexInputBindingDescription::inputRate must be of type VK_VERTEX_INPUT_RATE_INSTANCE for this binding.

See Also

VkPipelineVertexInputDivisorStateCreateInfoEXT

Constructors

VkVertexInputBindingDivisorDescriptionEXT

Fields

vkBinding :: Word32

binding is the binding number for which the divisor is specified.

vkDivisor :: Word32

divisor is the number of successive instances that will use the same value of the vertex attribute when instanced rendering is enabled. For example, if the divisor is N, the same vertex attribute will applied to N successive instances before moving on to the next vertex attribute. The maximum value of divisor is implementation dependent and can be queried using VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT::maxVertexAttribDivisor. A value of 0 can be used for the divisor if the vertexAttributeInstanceRateZeroDivisor feature is enabled. In this case, the same vertex attribute will be applied to all instances.

pattern VK_EXT_VERTEX_ATTRIBUTE_DIVISOR_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_EXT_VERTEX_ATTRIBUTE_DIVISOR_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VERTEX_ATTRIBUTE_DIVISOR_FEATURES_EXT :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VERTEX_ATTRIBUTE_DIVISOR_PROPERTIES_EXT :: VkStructureType

pattern VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_DIVISOR_STATE_CREATE_INFO_EXT :: VkStructureType