Graphics.Vulkan.C.Extensions.VK_FUCHSIA_imagepipe_surface

Documentation

newtype VkImagePipeSurfaceCreateFlagsFUCHSIA

Constructors

VkImagePipeSurfaceCreateFlagsFUCHSIA VkFlags

data VkImagePipeSurfaceCreateInfoFUCHSIA

VkImagePipeSurfaceCreateInfoFUCHSIA - Structure specifying parameters of a newly created ImagePipe surface object

Valid Usage (Implicit)

See Also

VkImagePipeSurfaceCreateFlagsFUCHSIA, VkStructureType, vkCreateImagePipeSurfaceFUCHSIA

Constructors

VkImagePipeSurfaceCreateInfoFUCHSIA

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_IMAGEPIPE_SURFACE_CREATE_INFO_FUCHSIA

vkPNext :: Ptr ()

pNext must be NULL

vkFlags :: VkImagePipeSurfaceCreateFlagsFUCHSIA

flags must be 0

vkImagePipeHandle :: Zx_handle_t

imagePipeHandle must be a valid zx_handle_t

type Zx_handle_t = Word32

type FN_vkCreateImagePipeSurfaceFUCHSIA = ("instance" ::: VkInstance) -> ("pCreateInfo" ::: Ptr VkImagePipeSurfaceCreateInfoFUCHSIA) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pSurface" ::: Ptr VkSurfaceKHR) -> IO VkResult

type PFN_vkCreateImagePipeSurfaceFUCHSIA = FunPtr FN_vkCreateImagePipeSurfaceFUCHSIA

vkCreateImagePipeSurfaceFUCHSIA :: InstanceCmds -> ("instance" ::: VkInstance) -> ("pCreateInfo" ::: Ptr VkImagePipeSurfaceCreateInfoFUCHSIA) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pSurface" ::: Ptr VkSurfaceKHR) -> IO VkResult

vkCreateImagePipeSurfaceFUCHSIA - Create a VkSurfaceKHR object for a Fuchsia ImagePipe

Parameters

instance is the instance to associate with the surface.

pCreateInfo is a pointer to an instance of the VkImagePipeSurfaceCreateInfoFUCHSIA structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkImagePipeSurfaceCreateInfoFUCHSIA structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkImagePipeSurfaceCreateInfoFUCHSIA, VkInstance, VkSurfaceKHR

pattern VK_FUCHSIA_IMAGEPIPE_SURFACE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_FUCHSIA_IMAGEPIPE_SURFACE_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_IMAGEPIPE_SURFACE_CREATE_INFO_FUCHSIA :: VkStructureType