Graphics.Vulkan.C.Extensions.VK_GGP_stream_descriptor_surface

Documentation

type GgpStreamDescriptor = Word32

newtype VkStreamDescriptorSurfaceCreateFlagsGGP

Constructors

VkStreamDescriptorSurfaceCreateFlagsGGP VkFlags

data VkStreamDescriptorSurfaceCreateInfoGGP

VkStreamDescriptorSurfaceCreateInfoGGP - Structure specifying parameters of a newly created Google Games Platform stream surface object

Valid Usage (Implicit)

See Also

VkStreamDescriptorSurfaceCreateFlagsGGP, VkStructureType, vkCreateStreamDescriptorSurfaceGGP

Constructors

VkStreamDescriptorSurfaceCreateInfoGGP

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_STREAM_DESCRIPTOR_SURFACE_CREATE_INFO_GGP

vkPNext :: Ptr ()

pNext must be NULL

vkFlags :: VkStreamDescriptorSurfaceCreateFlagsGGP

flags must be 0

vkStreamDescriptor :: GgpStreamDescriptor

streamDescriptor must be a valid GgpStreamDescriptor

type FN_vkCreateStreamDescriptorSurfaceGGP = ("instance" ::: VkInstance) -> ("pCreateInfo" ::: Ptr VkStreamDescriptorSurfaceCreateInfoGGP) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pSurface" ::: Ptr VkSurfaceKHR) -> IO VkResult

type PFN_vkCreateStreamDescriptorSurfaceGGP = FunPtr FN_vkCreateStreamDescriptorSurfaceGGP

vkCreateStreamDescriptorSurfaceGGP :: InstanceCmds -> ("instance" ::: VkInstance) -> ("pCreateInfo" ::: Ptr VkStreamDescriptorSurfaceCreateInfoGGP) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pSurface" ::: Ptr VkSurfaceKHR) -> IO VkResult

vkCreateStreamDescriptorSurfaceGGP - Create a VkSurfaceKHR object for a Google Games Platform stream

Parameters

instance is the instance to associate with the surface.

pCreateInfo is a pointer to an instance of the VkStreamDescriptorSurfaceCreateInfoGGP structure containing parameters that affect the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkStreamDescriptorSurfaceCreateInfoGGP structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_NATIVE_WINDOW_IN_USE_KHR

See Also

VkAllocationCallbacks, VkInstance, VkStreamDescriptorSurfaceCreateInfoGGP, VkSurfaceKHR

pattern VK_GGP_STREAM_DESCRIPTOR_SURFACE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_GGP_STREAM_DESCRIPTOR_SURFACE_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_STREAM_DESCRIPTOR_SURFACE_CREATE_INFO_GGP :: VkStructureType