Graphics.Vulkan.C.Extensions.VK_KHR_8bit_storage

Documentation

data VkPhysicalDevice8BitStorageFeaturesKHR

VkPhysicalDevice8BitStorageFeaturesKHR - Structure describing features supported by VK_KHR_8bit_storage

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

VkPhysicalDevice8BitStorageFeaturesKHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_8BIT_STORAGE_FEATURES_KHR

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkStorageBuffer8BitAccess :: VkBool32

storageBuffer8BitAccess indicates whether objects in the StorageBuffer storage class with the Block decoration can have 8-bit integer members. If this feature is not enabled, 8-bit integer members must not be used in such objects. This also indicates whether shader modules can declare the StorageBuffer8BitAccess capability.

vkUniformAndStorageBuffer8BitAccess :: VkBool32

uniformAndStorageBuffer8BitAccess indicates whether objects in the Uniform storage class with the Block decoration and in the StorageBuffer storage class with the same decoration can have 8-bit integer members. If this feature is not enabled, 8-bit integer members must not be used in such objects. This also indicates whether shader modules can declare the UniformAndStorageBuffer8BitAccess capability.

vkStoragePushConstant8 :: VkBool32

storagePushConstant8 indicates whether objects in the PushConstant storage class can have 8-bit integer members. If this feature is not enabled, 8-bit integer members must not be used in such objects. This also indicates whether shader modules can declare the StoragePushConstant8 capability.

pattern VK_KHR_8BIT_STORAGE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_8BIT_STORAGE_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_8BIT_STORAGE_FEATURES_KHR :: VkStructureType