Graphics.Vulkan.C.Extensions.VK_KHR_depth_stencil_resolve

Documentation

data VkPhysicalDeviceDepthStencilResolvePropertiesKHR

VkPhysicalDeviceDepthStencilResolvePropertiesKHR - Structure describing depth/stencil resolve properties that can be supported by an implementation

Members

The members of the VkPhysicalDeviceDepthStencilResolvePropertiesKHR structure describe the following implementation-dependent limits:

Valid Usage (Implicit)

See Also

VkBool32, VkResolveModeFlagsKHR, VkStructureType

Constructors

VkPhysicalDeviceDepthStencilResolvePropertiesKHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DEPTH_STENCIL_RESOLVE_PROPERTIES_KHR

vkPNext :: Ptr ()

vkSupportedDepthResolveModes :: VkResolveModeFlagsKHR

supportedDepthResolveModes is a bitmask of VkResolveModeFlagBitsKHR indicating the set of supported depth resolve modes. VK_RESOLVE_MODE_SAMPLE_ZERO_BIT_KHR must be included in the set but implementations may support additional modes.

vkSupportedStencilResolveModes :: VkResolveModeFlagsKHR

supportedStencilResolveModes is a bitmask of VkResolveModeFlagBitsKHR indicating the set of supported stencil resolve modes. VK_RESOLVE_MODE_SAMPLE_ZERO_BIT_KHR must be included in the set but implementations may support additional modes. VK_RESOLVE_MODE_AVERAGE_BIT_KHR must not be included in the set.

vkIndependentResolveNone :: VkBool32

independentResolveNone is VK_TRUE if the implementation supports setting the depth and stencil resolve modes to different values when one of those modes is VK_RESOLVE_MODE_NONE_KHR. Otherwise the implementation only supports setting both modes to the same value.

vkIndependentResolve :: VkBool32

independentResolve is VK_TRUE if the implementation supports all combinations of the supported depth and stencil resolve modes.

newtype VkResolveModeFlagBitsKHR

VkResolveModeFlagBitsKHR - Bitmask indicating supported depth and stencil resolve modes

See Also

VkResolveModeFlagsKHR, VkSubpassDescriptionDepthStencilResolveKHR

Constructors

VkResolveModeFlagBitsKHR VkFlags

pattern VK_RESOLVE_MODE_SAMPLE_ZERO_BIT_KHR :: VkResolveModeFlagBitsKHR

VK_RESOLVE_MODE_SAMPLE_ZERO_BIT_KHR indicates that result of the resolve operation is equal to the value of sample 0.

pattern VK_RESOLVE_MODE_AVERAGE_BIT_KHR :: VkResolveModeFlagBitsKHR

VK_RESOLVE_MODE_AVERAGE_BIT_KHR indicates that result of the resolve operation is the average of the sample values.

pattern VK_RESOLVE_MODE_MIN_BIT_KHR :: VkResolveModeFlagBitsKHR

VK_RESOLVE_MODE_MIN_BIT_KHR indicates that result of the resolve operation is the minimum of the sample values.

pattern VK_RESOLVE_MODE_MAX_BIT_KHR :: VkResolveModeFlagBitsKHR

VK_RESOLVE_MODE_MAX_BIT_KHR indicates that result of the resolve operation is the maximum of the sample values.

pattern VK_RESOLVE_MODE_NONE_KHR :: VkResolveModeFlagBitsKHR

VK_RESOLVE_MODE_NONE_KHR indicates that no resolve operation is done.

type VkResolveModeFlagsKHR = VkResolveModeFlagBitsKHR

VkResolveModeFlagsKHR - Bitmask of VkResolveModeFlagBitsKHR

Description

VkResolveModeFlagsKHR is a bitmask type for setting a mask of zero or more VkResolveModeFlagBitsKHR.

See Also

VkPhysicalDeviceDepthStencilResolvePropertiesKHR, VkResolveModeFlagBitsKHR

data VkSubpassDescriptionDepthStencilResolveKHR

VkSubpassDescriptionDepthStencilResolveKHR - Structure specifying depth/stencil resolve operations for a subpass

Valid Usage

If pDepthStencilResolveAttachment is not NULL and does not have the value VK_ATTACHMENT_UNUSED, pDepthStencilAttachment must not have the value VK_ATTACHMENT_UNUSED

If pDepthStencilResolveAttachment is not NULL and does not have the value VK_ATTACHMENT_UNUSED, depthResolveMode and stencilResolveMode must not both be VK_RESOLVE_MODE_NONE_KHR

If pDepthStencilResolveAttachment is not NULL and does not have the value VK_ATTACHMENT_UNUSED, pDepthStencilAttachment must not have a sample count of VK_SAMPLE_COUNT_1_BIT

If pDepthStencilResolveAttachment is not NULL and does not have the value VK_ATTACHMENT_UNUSED, pDepthStencilResolveAttachment must have a sample count of VK_SAMPLE_COUNT_1_BIT

If pDepthStencilResolveAttachment is not NULL and does not have the value VK_ATTACHMENT_UNUSED then it must have a format whose features contain VK_FORMAT_FEATURE_DEPTH_STENCIL_ATTACHMENT_BIT

If the VkFormat of pDepthStencilResolveAttachment has a depth component, then the VkFormat of pDepthStencilAttachment must have a depth component with the same number of bits and numerical type

If the VkFormat of pDepthStencilResolveAttachment has a stencil component, then the VkFormat of pDepthStencilAttachment must have a stencil component with the same number of bits and numerical type

The value of depthResolveMode must be one of the bits set in VkPhysicalDeviceDepthStencilResolvePropertiesKHR::supportedDepthResolveModes or VK_RESOLVE_MODE_NONE_KHR

The value of stencilResolveMode must be one of the bits set in VkPhysicalDeviceDepthStencilResolvePropertiesKHR::supportedStencilResolveModes or VK_RESOLVE_MODE_NONE_KHR

If the VkFormat of pDepthStencilResolveAttachment has both depth and stencil components, VkPhysicalDeviceDepthStencilResolvePropertiesKHR::independentResolve is VK_FALSE, and VkPhysicalDeviceDepthStencilResolvePropertiesKHR::independentResolveNone is VK_FALSE, then the values of depthResolveMode and stencilResolveMode must be identical

If the VkFormat of pDepthStencilResolveAttachment has both depth and stencil components, VkPhysicalDeviceDepthStencilResolvePropertiesKHR::independentResolve is VK_FALSE and VkPhysicalDeviceDepthStencilResolvePropertiesKHR::independentResolveNone is VK_TRUE, then the values of depthResolveMode and stencilResolveMode must be identical or one of them must be VK_RESOLVE_MODE_NONE_KHR

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_SUBPASS_DESCRIPTION_DEPTH_STENCIL_RESOLVE_KHR

depthResolveMode must be a valid VkResolveModeFlagBitsKHR value

stencilResolveMode must be a valid VkResolveModeFlagBitsKHR value

If pDepthStencilResolveAttachment is not NULL, pDepthStencilResolveAttachment must be a valid pointer to a valid VkAttachmentReference2KHR structure

See Also

VkAttachmentReference2KHR, VkResolveModeFlagBitsKHR, VkStructureType

Constructors

VkSubpassDescriptionDepthStencilResolveKHR

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkDepthResolveMode :: VkResolveModeFlagBitsKHR

depthResolveMode is a bitmask of VkResolveModeFlagBitsKHR describing the depth resolve mode.

vkStencilResolveMode :: VkResolveModeFlagBitsKHR

stencilResolveMode is a bitmask of VkResolveModeFlagBitsKHR describing the stencil resolve mode.

vkPDepthStencilResolveAttachment :: Ptr VkAttachmentReference2KHR

pDepthStencilResolveAttachment is an optional VkAttachmentReference structure defining the depth/stencil resolve attachment for this subpass and its layout.

pattern VK_KHR_DEPTH_STENCIL_RESOLVE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_DEPTH_STENCIL_RESOLVE_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DEPTH_STENCIL_RESOLVE_PROPERTIES_KHR :: VkStructureType

pattern VK_STRUCTURE_TYPE_SUBPASS_DESCRIPTION_DEPTH_STENCIL_RESOLVE_KHR :: VkStructureType