Graphics.Vulkan.C.Extensions.VK_KHR_display

Documentation

type VkDisplayKHR = Ptr VkDisplayKHR_T

VkDisplayKHR - Opaque handle to a display object

See Also

VkDisplayPlanePropertiesKHR, VkDisplayPropertiesKHR, vkAcquireXlibDisplayEXT, vkCreateDisplayModeKHR, vkDisplayPowerControlEXT, vkGetDisplayModeProperties2KHR, vkGetDisplayModePropertiesKHR, vkGetDisplayPlaneSupportedDisplaysKHR, vkGetRandROutputDisplayEXT, vkRegisterDisplayEventEXT, vkReleaseDisplayEXT

newtype VkDisplayModeCreateFlagsKHR

Constructors

VkDisplayModeCreateFlagsKHR VkFlags

data VkDisplayModeCreateInfoKHR

VkDisplayModeCreateInfoKHR - Structure specifying parameters of a newly created display mode object

Valid Usage (Implicit)

See Also

VkDisplayModeCreateFlagsKHR, VkDisplayModeParametersKHR, VkStructureType, vkCreateDisplayModeKHR

Constructors

VkDisplayModeCreateInfoKHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DISPLAY_MODE_CREATE_INFO_KHR

vkPNext :: Ptr ()

pNext must be NULL

vkFlags :: VkDisplayModeCreateFlagsKHR

flags must be 0

vkParameters :: VkDisplayModeParametersKHR

parameters must be a valid VkDisplayModeParametersKHR structure

type VkDisplayModeKHR = Ptr VkDisplayModeKHR_T

VkDisplayModeKHR - Opaque handle to a display mode object

See Also

VkDisplayModePropertiesKHR, VkDisplayPlaneInfo2KHR, VkDisplaySurfaceCreateInfoKHR, vkCreateDisplayModeKHR, vkGetDisplayPlaneCapabilitiesKHR

data VkDisplayModeParametersKHR

VkDisplayModeParametersKHR - Structure describing display parameters associated with a display mode

Description

Note

For example, a 60Hz display mode would report a refreshRate of 60,000.

Valid Usage

The width member of visibleRegion must be greater than 0

The height member of visibleRegion must be greater than 0

refreshRate must be greater than 0

See Also

VkDisplayModeCreateInfoKHR, VkDisplayModePropertiesKHR, VkExtent2D

Constructors

VkDisplayModeParametersKHR

Fields

vkVisibleRegion :: VkExtent2D

visibleRegion is the 2D extents of the visible region.

vkRefreshRate :: Word32

refreshRate is a uint32_t that is the number of times the display is refreshed each second multiplied by 1000.

data VkDisplayModePropertiesKHR

VkDisplayModePropertiesKHR - Structure describing display mode properties

See Also

VkDisplayModeKHR, VkDisplayModeParametersKHR, VkDisplayModeProperties2KHR, vkGetDisplayModePropertiesKHR

Constructors

VkDisplayModePropertiesKHR

Fields

vkDisplayMode :: VkDisplayModeKHR

displayMode is a handle to the display mode described in this structure. This handle will be valid for the lifetime of the Vulkan instance.

vkParameters :: VkDisplayModeParametersKHR

parameters is a VkDisplayModeParametersKHR structure describing the display parameters associated with displayMode.

newtype VkDisplayPlaneAlphaFlagBitsKHR

VkDisplayPlaneAlphaFlagBitsKHR - Alpha blending type

See Also

VkDisplayPlaneAlphaFlagsKHR, VkDisplaySurfaceCreateInfoKHR

Constructors

VkDisplayPlaneAlphaFlagBitsKHR VkFlags

pattern VK_DISPLAY_PLANE_ALPHA_OPAQUE_BIT_KHR :: VkDisplayPlaneAlphaFlagBitsKHR

VK_DISPLAY_PLANE_ALPHA_OPAQUE_BIT_KHR specifies that the source image will be treated as opaque.

pattern VK_DISPLAY_PLANE_ALPHA_GLOBAL_BIT_KHR :: VkDisplayPlaneAlphaFlagBitsKHR

VK_DISPLAY_PLANE_ALPHA_GLOBAL_BIT_KHR specifies that a global alpha value must be specified that will be applied to all pixels in the source image.

pattern VK_DISPLAY_PLANE_ALPHA_PER_PIXEL_BIT_KHR :: VkDisplayPlaneAlphaFlagBitsKHR

VK_DISPLAY_PLANE_ALPHA_PER_PIXEL_BIT_KHR specifies that the alpha value will be determined by the alpha channel of the source image’s pixels. If the source format contains no alpha values, no blending will be applied. The source alpha values are not premultiplied into the source image’s other color channels.

pattern VK_DISPLAY_PLANE_ALPHA_PER_PIXEL_PREMULTIPLIED_BIT_KHR :: VkDisplayPlaneAlphaFlagBitsKHR

VK_DISPLAY_PLANE_ALPHA_PER_PIXEL_PREMULTIPLIED_BIT_KHR is equivalent to VK_DISPLAY_PLANE_ALPHA_PER_PIXEL_BIT_KHR, except the source alpha values are assumed to be premultiplied into the source image’s other color channels.

type VkDisplayPlaneAlphaFlagsKHR = VkDisplayPlaneAlphaFlagBitsKHR

VkDisplayPlaneAlphaFlagsKHR - Bitmask of VkDisplayPlaneAlphaFlagBitsKHR

Description

VkDisplayPlaneAlphaFlagsKHR is a bitmask type for setting a mask of zero or more VkDisplayPlaneAlphaFlagBitsKHR.

See Also

VkDisplayPlaneAlphaFlagBitsKHR, VkDisplayPlaneCapabilitiesKHR

data VkDisplayPlaneCapabilitiesKHR

VkDisplayPlaneCapabilitiesKHR - Structure describing capabilities of a mode and plane combination

Description

The minimum and maximum position and extent fields describe the implementation limits, if any, as they apply to the specified display mode and plane. Vendors may support displaying a subset of a swapchain’s presentable images on the specified display plane. This is expressed by returning minSrcPosition, maxSrcPosition, minSrcExtent, and maxSrcExtent values that indicate a range of possible positions and sizes may be used to specify the region within the presentable images that source pixels will be read from when creating a swapchain on the specified display mode and plane.

Vendors may also support mapping the presentable images’ content to a subset or superset of the visible region in the specified display mode. This is expressed by returning minDstPosition, maxDstPosition, minDstExtent and maxDstExtent values that indicate a range of possible positions and sizes may be used to describe the region within the display mode that the source pixels will be mapped to.

Other vendors may support only a 1-1 mapping between pixels in the presentable images and the display mode. This may be indicated by returning (0,0) for minSrcPosition, maxSrcPosition, minDstPosition, and maxDstPosition, and (display mode width, display mode height) for minSrcExtent, maxSrcExtent, minDstExtent, and maxDstExtent.

These values indicate the limits of the implementation’s individual fields. Not all combinations of values within the offset and extent ranges returned in VkDisplayPlaneCapabilitiesKHR are guaranteed to be supported. Vendors may still fail presentation requests that specify unsupported combinations.

See Also

VkDisplayPlaneAlphaFlagsKHR, VkDisplayPlaneCapabilities2KHR, VkExtent2D, VkOffset2D, vkGetDisplayPlaneCapabilitiesKHR

Constructors

VkDisplayPlaneCapabilitiesKHR

Fields

vkSupportedAlpha :: VkDisplayPlaneAlphaFlagsKHR

supportedAlpha is a bitmask of VkDisplayPlaneAlphaFlagBitsKHR describing the supported alpha blending modes.

vkMinSrcPosition :: VkOffset2D

minSrcPosition is the minimum source rectangle offset supported by this plane using the specified mode.

vkMaxSrcPosition :: VkOffset2D

maxSrcPosition is the maximum source rectangle offset supported by this plane using the specified mode. The x and y components of maxSrcPosition must each be greater than or equal to the x and y components of minSrcPosition, respectively.

vkMinSrcExtent :: VkExtent2D

minSrcExtent is the minimum source rectangle size supported by this plane using the specified mode.

vkMaxSrcExtent :: VkExtent2D

maxSrcExtent is the maximum source rectangle size supported by this plane using the specified mode.

vkMinDstPosition :: VkOffset2D

minDstPosition, maxDstPosition, minDstExtent, maxDstExtent all have similar semantics to their corresponding *Src* equivalents, but apply to the output region within the mode rather than the input region within the source image. Unlike the *Src* offsets, minDstPosition and maxDstPosition may contain negative values.

vkMaxDstPosition :: VkOffset2D

vkMinDstExtent :: VkExtent2D

vkMaxDstExtent :: VkExtent2D

data VkDisplayPlanePropertiesKHR

VkDisplayPlanePropertiesKHR - Structure describing display plane properties

See Also

VkDisplayKHR, VkDisplayPlaneProperties2KHR, vkGetPhysicalDeviceDisplayPlanePropertiesKHR

Constructors

VkDisplayPlanePropertiesKHR

Fields

vkCurrentDisplay :: VkDisplayKHR

currentDisplay is the handle of the display the plane is currently associated with. If the plane is not currently attached to any displays, this will be VK_NULL_HANDLE.

vkCurrentStackIndex :: Word32

currentStackIndex is the current z-order of the plane. This will be between 0 and the value returned by vkGetPhysicalDeviceDisplayPlanePropertiesKHR in pPropertyCount.

data VkDisplayPropertiesKHR

VkDisplayPropertiesKHR - Structure describing an available display device

Description

Note

For devices which have no natural value to return here, implementations should return the maximum resolution supported.

Note

Persistent presents may have higher latency, and may use less power when the screen content is updated infrequently, or when only a portion of the screen needs to be updated in most frames.

See Also

VkBool32, VkDisplayKHR, VkDisplayProperties2KHR, VkExtent2D, VkSurfaceTransformFlagsKHR, vkGetPhysicalDeviceDisplayPropertiesKHR

Constructors

VkDisplayPropertiesKHR

Fields

vkDisplay :: VkDisplayKHR

display is a handle that is used to refer to the display described here. This handle will be valid for the lifetime of the Vulkan instance.

vkDisplayName :: Ptr CChar

displayName is a pointer to a NULL-terminated string containing the name of the display. Generally, this will be the name provided by the display’s EDID. It can be NULL if no suitable name is available. If not NULL, the memory it points to must remain accessible as long as display is valid.

vkPhysicalDimensions :: VkExtent2D

physicalDimensions describes the physical width and height of the visible portion of the display, in millimeters.

vkPhysicalResolution :: VkExtent2D

physicalResolution describes the physical, native, or preferred resolution of the display.

vkSupportedTransforms :: VkSurfaceTransformFlagsKHR

supportedTransforms is a bitmask of VkSurfaceTransformFlagBitsKHR describing which transforms are supported by this display.

vkPlaneReorderPossible :: VkBool32

planeReorderPossible tells whether the planes on this display can have their z order changed. If this is VK_TRUE, the application can re-arrange the planes on this display in any order relative to each other.

vkPersistentContent :: VkBool32

persistentContent tells whether the display supports self-refresh/internal buffering. If this is true, the application can submit persistent present operations on swapchains created against this display.

newtype VkDisplaySurfaceCreateFlagsKHR

Constructors

VkDisplaySurfaceCreateFlagsKHR VkFlags

data VkDisplaySurfaceCreateInfoKHR

VkDisplaySurfaceCreateInfoKHR - Structure specifying parameters of a newly created display plane surface object

Description

Note

Creating a display surface must not modify the state of the displays, planes, or other resources it names. For example, it must not apply the specified mode to be set on the associated display. Application of display configuration occurs as a side effect of presenting to a display surface.

Valid Usage

planeIndex must be less than the number of display planes supported by the device as determined by calling vkGetPhysicalDeviceDisplayPlanePropertiesKHR

If the planeReorderPossible member of the VkDisplayPropertiesKHR structure returned by vkGetPhysicalDeviceDisplayPropertiesKHR for the display corresponding to displayMode is VK_TRUE then planeStackIndex must be less than the number of display planes supported by the device as determined by calling vkGetPhysicalDeviceDisplayPlanePropertiesKHR; otherwise planeStackIndex must equal the currentStackIndex member of VkDisplayPlanePropertiesKHR returned by vkGetPhysicalDeviceDisplayPlanePropertiesKHR for the display plane corresponding to displayMode

If alphaMode is VK_DISPLAY_PLANE_ALPHA_GLOBAL_BIT_KHR then globalAlpha must be between 0 and 1, inclusive

alphaMode must be 0 or one of the bits present in the supportedAlpha member of VkDisplayPlaneCapabilitiesKHR returned by vkGetDisplayPlaneCapabilitiesKHR for the display plane corresponding to displayMode

The width and height members of imageExtent must be less than the maxImageDimensions2D member of VkPhysicalDeviceLimits

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DISPLAY_SURFACE_CREATE_INFO_KHR

pNext must be NULL

flags must be 0

displayMode must be a valid VkDisplayModeKHR handle

transform must be a valid VkSurfaceTransformFlagBitsKHR value

alphaMode must be a valid VkDisplayPlaneAlphaFlagBitsKHR value

See Also

VkDisplayModeKHR, VkDisplayPlaneAlphaFlagBitsKHR, VkDisplaySurfaceCreateFlagsKHR, VkExtent2D, VkStructureType, VkSurfaceTransformFlagBitsKHR, vkCreateDisplayPlaneSurfaceKHR

Constructors

VkDisplaySurfaceCreateInfoKHR

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkFlags :: VkDisplaySurfaceCreateFlagsKHR

flags is reserved for future use, and must be zero.

vkDisplayMode :: VkDisplayModeKHR

displayMode is a VkDisplayModeKHR handle specifying the mode to use when displaying this surface.

vkPlaneIndex :: Word32

planeIndex is the plane on which this surface appears.

vkPlaneStackIndex :: Word32

planeStackIndex is the z-order of the plane.

vkTransform :: VkSurfaceTransformFlagBitsKHR

transform is a VkSurfaceTransformFlagBitsKHR value specifying the transformation to apply to images as part of the scanout operation.

vkGlobalAlpha :: CFloat

globalAlpha is the global alpha value. This value is ignored if alphaMode is not VK_DISPLAY_PLANE_ALPHA_GLOBAL_BIT_KHR.

vkAlphaMode :: VkDisplayPlaneAlphaFlagBitsKHR

alphaMode is a VkDisplayPlaneAlphaFlagBitsKHR value specifying the type of alpha blending to use.

vkImageExtent :: VkExtent2D

imageExtent The size of the presentable images to use with the surface.

type FN_vkCreateDisplayModeKHR = ("physicalDevice" ::: VkPhysicalDevice) -> ("display" ::: VkDisplayKHR) -> ("pCreateInfo" ::: Ptr VkDisplayModeCreateInfoKHR) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pMode" ::: Ptr VkDisplayModeKHR) -> IO VkResult

type PFN_vkCreateDisplayModeKHR = FunPtr FN_vkCreateDisplayModeKHR

vkCreateDisplayModeKHR :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("display" ::: VkDisplayKHR) -> ("pCreateInfo" ::: Ptr VkDisplayModeCreateInfoKHR) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pMode" ::: Ptr VkDisplayModeKHR) -> IO VkResult

vkCreateDisplayModeKHR - Create a display mode

Parameters

physicalDevice is the physical device associated with display.

display is the display to create an additional mode for.

pCreateInfo is a VkDisplayModeCreateInfoKHR structure describing the new mode to create.

pAllocator is the allocator used for host memory allocated for the display mode object when there is no more specific allocator available (see Memory Allocation).

pMode returns the handle of the mode created.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

display must be a valid VkDisplayKHR handle

pCreateInfo must be a valid pointer to a valid VkDisplayModeCreateInfoKHR structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pMode must be a valid pointer to a VkDisplayModeKHR handle

Host Synchronization

Host access to display must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_INITIALIZATION_FAILED

See Also

VkAllocationCallbacks, VkDisplayKHR, VkDisplayModeCreateInfoKHR, VkDisplayModeKHR, VkPhysicalDevice

type FN_vkCreateDisplayPlaneSurfaceKHR = ("instance" ::: VkInstance) -> ("pCreateInfo" ::: Ptr VkDisplaySurfaceCreateInfoKHR) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pSurface" ::: Ptr VkSurfaceKHR) -> IO VkResult

type PFN_vkCreateDisplayPlaneSurfaceKHR = FunPtr FN_vkCreateDisplayPlaneSurfaceKHR

vkCreateDisplayPlaneSurfaceKHR :: InstanceCmds -> ("instance" ::: VkInstance) -> ("pCreateInfo" ::: Ptr VkDisplaySurfaceCreateInfoKHR) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pSurface" ::: Ptr VkSurfaceKHR) -> IO VkResult

vkCreateDisplayPlaneSurfaceKHR - Create a VkSurfaceKHR structure representing a display plane and mode

Parameters

instance is the instance corresponding to the physical device the targeted display is on.

pCreateInfo is a pointer to an instance of the VkDisplaySurfaceCreateInfoKHR structure specifying which mode, plane, and other parameters to use, as described below.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface is returned.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkDisplaySurfaceCreateInfoKHR structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDisplaySurfaceCreateInfoKHR, VkInstance, VkSurfaceKHR

type FN_vkGetDisplayModePropertiesKHR = ("physicalDevice" ::: VkPhysicalDevice) -> ("display" ::: VkDisplayKHR) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkDisplayModePropertiesKHR) -> IO VkResult

type PFN_vkGetDisplayModePropertiesKHR = FunPtr FN_vkGetDisplayModePropertiesKHR

vkGetDisplayModePropertiesKHR :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("display" ::: VkDisplayKHR) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkDisplayModePropertiesKHR) -> IO VkResult

vkGetDisplayModePropertiesKHR - Query the set of mode properties supported by the display

Parameters

physicalDevice is the physical device associated with display.

display is the display to query.

pPropertyCount is a pointer to an integer related to the number of display modes available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayModePropertiesKHR structures.

Description

If pProperties is NULL, then the number of display modes available on the specified display for physicalDevice is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If the value of pPropertyCount is less than the number of display modes for physicalDevice, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of display modes available on the specified display for physicalDevice, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

display must be a valid VkDisplayKHR handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayModePropertiesKHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayKHR, VkDisplayModePropertiesKHR, VkPhysicalDevice

type FN_vkGetDisplayPlaneCapabilitiesKHR = ("physicalDevice" ::: VkPhysicalDevice) -> ("mode" ::: VkDisplayModeKHR) -> ("planeIndex" ::: Word32) -> ("pCapabilities" ::: Ptr VkDisplayPlaneCapabilitiesKHR) -> IO VkResult

type PFN_vkGetDisplayPlaneCapabilitiesKHR = FunPtr FN_vkGetDisplayPlaneCapabilitiesKHR

vkGetDisplayPlaneCapabilitiesKHR :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("mode" ::: VkDisplayModeKHR) -> ("planeIndex" ::: Word32) -> ("pCapabilities" ::: Ptr VkDisplayPlaneCapabilitiesKHR) -> IO VkResult

vkGetDisplayPlaneCapabilitiesKHR - Query capabilities of a mode and plane combination

Parameters

physicalDevice is the physical device associated with display

mode is the display mode the application intends to program when using the specified plane. Note this parameter also implicitly specifies a display.

planeIndex is the plane which the application intends to use with the display, and is less than the number of display planes supported by the device.

pCapabilities is a pointer to a VkDisplayPlaneCapabilitiesKHR structure in which the capabilities are returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

mode must be a valid VkDisplayModeKHR handle

pCapabilities must be a valid pointer to a VkDisplayPlaneCapabilitiesKHR structure

Host Synchronization

Host access to mode must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayModeKHR, VkDisplayPlaneCapabilitiesKHR, VkPhysicalDevice

type FN_vkGetDisplayPlaneSupportedDisplaysKHR = ("physicalDevice" ::: VkPhysicalDevice) -> ("planeIndex" ::: Word32) -> ("pDisplayCount" ::: Ptr Word32) -> ("pDisplays" ::: Ptr VkDisplayKHR) -> IO VkResult

type PFN_vkGetDisplayPlaneSupportedDisplaysKHR = FunPtr FN_vkGetDisplayPlaneSupportedDisplaysKHR

vkGetDisplayPlaneSupportedDisplaysKHR :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("planeIndex" ::: Word32) -> ("pDisplayCount" ::: Ptr Word32) -> ("pDisplays" ::: Ptr VkDisplayKHR) -> IO VkResult

vkGetDisplayPlaneSupportedDisplaysKHR - Query the list of displays a plane supports

Parameters

physicalDevice is a physical device.

planeIndex is the plane which the application wishes to use, and must be in the range [0, physical device plane count - 1].

pDisplayCount is a pointer to an integer related to the number of displays available or queried, as described below.

pDisplays is either NULL or a pointer to an array of VkDisplayKHR handles.

Description

If pDisplays is NULL, then the number of displays usable with the specified planeIndex for physicalDevice is returned in pDisplayCount. Otherwise, pDisplayCount must point to a variable set by the user to the number of elements in the pDisplays array, and on return the variable is overwritten with the number of handles actually written to pDisplays. If the value of pDisplayCount is less than the number of display planes for physicalDevice, at most pDisplayCount handles will be written. If pDisplayCount is smaller than the number of displays usable with the specified planeIndex for physicalDevice, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage

planeIndex must be less than the number of display planes supported by the device as determined by calling vkGetPhysicalDeviceDisplayPlanePropertiesKHR

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pDisplayCount must be a valid pointer to a uint32_t value

If the value referenced by pDisplayCount is not 0, and pDisplays is not NULL, pDisplays must be a valid pointer to an array of pDisplayCount VkDisplayKHR handles

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayKHR, VkPhysicalDevice

type FN_vkGetPhysicalDeviceDisplayPlanePropertiesKHR = ("physicalDevice" ::: VkPhysicalDevice) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkDisplayPlanePropertiesKHR) -> IO VkResult

type PFN_vkGetPhysicalDeviceDisplayPlanePropertiesKHR = FunPtr FN_vkGetPhysicalDeviceDisplayPlanePropertiesKHR

vkGetPhysicalDeviceDisplayPlanePropertiesKHR :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkDisplayPlanePropertiesKHR) -> IO VkResult

vkGetPhysicalDeviceDisplayPlanePropertiesKHR - Query the plane properties

Parameters

physicalDevice is a physical device.

pPropertyCount is a pointer to an integer related to the number of display planes available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayPlanePropertiesKHR structures.

Description

If pProperties is NULL, then the number of display planes available for physicalDevice is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If the value of pPropertyCount is less than the number of display planes for physicalDevice, at most pPropertyCount structures will be written.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayPlanePropertiesKHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayPlanePropertiesKHR, VkPhysicalDevice

type FN_vkGetPhysicalDeviceDisplayPropertiesKHR = ("physicalDevice" ::: VkPhysicalDevice) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkDisplayPropertiesKHR) -> IO VkResult

type PFN_vkGetPhysicalDeviceDisplayPropertiesKHR = FunPtr FN_vkGetPhysicalDeviceDisplayPropertiesKHR

vkGetPhysicalDeviceDisplayPropertiesKHR :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkDisplayPropertiesKHR) -> IO VkResult

vkGetPhysicalDeviceDisplayPropertiesKHR - Query information about the available displays

Parameters

physicalDevice is a physical device.

pPropertyCount is a pointer to an integer related to the number of display devices available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayPropertiesKHR structures.

Description

If pProperties is NULL, then the number of display devices available for physicalDevice is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If the value of pPropertyCount is less than the number of display devices for physicalDevice, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of display devices available for physicalDevice, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayPropertiesKHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayPropertiesKHR, VkPhysicalDevice

pattern VK_KHR_DISPLAY_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_DISPLAY_SPEC_VERSION :: Integral a => a

pattern VK_OBJECT_TYPE_DISPLAY_KHR :: VkObjectType

pattern VK_OBJECT_TYPE_DISPLAY_MODE_KHR :: VkObjectType

pattern VK_STRUCTURE_TYPE_DISPLAY_MODE_CREATE_INFO_KHR :: VkStructureType

pattern VK_STRUCTURE_TYPE_DISPLAY_SURFACE_CREATE_INFO_KHR :: VkStructureType