Graphics.Vulkan.C.Extensions.VK_KHR_driver_properties

Documentation

data VkConformanceVersionKHR

VkConformanceVersionKHR - Structure containing the conformance test suite version the implementation is compliant with

See Also

VkPhysicalDeviceDriverPropertiesKHR

Constructors

VkConformanceVersionKHR

Fields

vkMajor :: Word8

major is the major version number of the conformance test suite.

vkMinor :: Word8

minor is the minor version number of the conformance test suite.

vkSubminor :: Word8

subminor is the subminor version number of the conformance test suite.

vkPatch :: Word8

patch is the patch version number of the conformance test suite.

newtype VkDriverIdKHR

VkDriverIdKHR - Khronos driver IDs

Description

Note

Khronos driver IDs may be allocated by vendors at any time. There may be multiple driver IDs for the same vendor, representing different drivers (for e.g. different platforms, proprietary or open source, etc.). Only the latest canonical versions of this Specification, of the corresponding vk.xml API Registry, and of the corresponding vulkan_core.h header file must contain all reserved Khronos driver IDs.

Only driver IDs registered with Khronos are given symbolic names. There may be unregistered driver IDs returned.

See Also

VkPhysicalDeviceDriverPropertiesKHR

Constructors

VkDriverIdKHR Int32

pattern VK_DRIVER_ID_AMD_PROPRIETARY_KHR :: VkDriverIdKHR

pattern VK_DRIVER_ID_AMD_OPEN_SOURCE_KHR :: VkDriverIdKHR

pattern VK_DRIVER_ID_MESA_RADV_KHR :: VkDriverIdKHR

pattern VK_DRIVER_ID_NVIDIA_PROPRIETARY_KHR :: VkDriverIdKHR

pattern VK_DRIVER_ID_INTEL_PROPRIETARY_WINDOWS_KHR :: VkDriverIdKHR

pattern VK_DRIVER_ID_INTEL_OPEN_SOURCE_MESA_KHR :: VkDriverIdKHR

pattern VK_DRIVER_ID_IMAGINATION_PROPRIETARY_KHR :: VkDriverIdKHR

pattern VK_DRIVER_ID_QUALCOMM_PROPRIETARY_KHR :: VkDriverIdKHR

pattern VK_DRIVER_ID_ARM_PROPRIETARY_KHR :: VkDriverIdKHR

pattern VK_DRIVER_ID_GOOGLE_PASTEL_KHR :: VkDriverIdKHR

pattern VK_DRIVER_ID_GGP_PROPRIETARY_KHR :: VkDriverIdKHR

data VkPhysicalDeviceDriverPropertiesKHR

VkPhysicalDeviceDriverPropertiesKHR - Structure containing driver identification information

Description

driverID must be immutable for a given driver across instances, processes, driver versions, and system reboots.

Valid Usage (Implicit)

See Also

VkConformanceVersionKHR, VkDriverIdKHR, VkStructureType

Constructors

VkPhysicalDeviceDriverPropertiesKHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DRIVER_PROPERTIES_KHR

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension specific structure.

vkDriverID :: VkDriverIdKHR

driverID is a unique identifier for the driver of the physical device.

vkDriverName :: Vector VK_MAX_DRIVER_NAME_SIZE_KHR CChar

driverName is a null-terminated UTF-8 string containing the name of the driver.

vkDriverInfo :: Vector VK_MAX_DRIVER_INFO_SIZE_KHR CChar

driverInfo is a null-terminated UTF-8 string containing additional information about the driver.

vkConformanceVersion :: VkConformanceVersionKHR

conformanceVersion is the version of the Vulkan conformance test this driver is conformant against (see VkConformanceVersionKHR).

pattern VK_KHR_DRIVER_PROPERTIES_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_DRIVER_PROPERTIES_SPEC_VERSION :: Integral a => a

type VK_MAX_DRIVER_INFO_SIZE_KHR = 256

pattern VK_MAX_DRIVER_INFO_SIZE_KHR :: Integral a => a

type VK_MAX_DRIVER_NAME_SIZE_KHR = 256

pattern VK_MAX_DRIVER_NAME_SIZE_KHR :: Integral a => a

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DRIVER_PROPERTIES_KHR :: VkStructureType