Graphics.Vulkan.C.Extensions.VK_KHR_external_fence

Documentation

type VkExportFenceCreateInfoKHR = VkExportFenceCreateInfo

pattern VkExportFenceCreateInfoKHR :: ("sType" ::: VkStructureType) -> ("pNext" ::: Ptr ()) -> ("handleTypes" ::: VkExternalFenceHandleTypeFlags) -> VkExportFenceCreateInfoKHR

type VkFenceImportFlagBitsKHR = VkFenceImportFlagBits

type VkFenceImportFlagsKHR = VkFenceImportFlags

pattern VK_FENCE_IMPORT_TEMPORARY_BIT_KHR :: VkFenceImportFlagBits

pattern VK_KHR_EXTERNAL_FENCE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_EXTERNAL_FENCE_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_EXPORT_FENCE_CREATE_INFO_KHR :: VkStructureType

pattern VK_STRUCTURE_TYPE_EXPORT_FENCE_CREATE_INFO :: VkStructureType

pattern VK_FENCE_IMPORT_TEMPORARY_BIT :: VkFenceImportFlagBits

VK_FENCE_IMPORT_TEMPORARY_BIT specifies that the fence payload will be imported only temporarily, as described in Importing Fence Payloads, regardless of the permanence of handleType.