Graphics.Vulkan.C.Extensions.VK_KHR_external_fence_win32

Documentation

data VkExportFenceWin32HandleInfoKHR

VkExportFenceWin32HandleInfoKHR - Structure specifying additional attributes of Windows handles exported from a fence

Description

If this structure is not present, or if pAttributes is set to NULL, default security descriptor values will be used, and child processes created by the application will not inherit the handle, as described in the MSDN documentation for “Synchronization Object Security and Access Rights”1. Further, if the structure is not present, the access rights will be

DXGI_SHARED_RESOURCE_READ | DXGI_SHARED_RESOURCE_WRITE

for handles of the following types:

VK_EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_BIT

1

https://msdn.microsoft.com/en-us/library/windows/desktop/ms686670.aspx

Valid Usage

If VkExportFenceCreateInfo::handleTypes does not include VK_EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_BIT, VkExportFenceWin32HandleInfoKHR must not be in the pNext chain of VkFenceCreateInfo.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_EXPORT_FENCE_WIN32_HANDLE_INFO_KHR

If pAttributes is not NULL, pAttributes must be a valid pointer to a valid SECURITY_ATTRIBUTES value

See Also

VkStructureType

Constructors

VkExportFenceWin32HandleInfoKHR

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkPAttributes :: Ptr SECURITY_ATTRIBUTES

pAttributes is a pointer to a Windows SECURITY_ATTRIBUTES structure specifying security attributes of the handle.

vkDwAccess :: DWORD

dwAccess is a DWORD specifying access rights of the handle.

vkName :: LPCWSTR

name is a NULL-terminated UTF-16 string to associate with the underlying synchronization primitive referenced by NT handles exported from the created fence.

data VkFenceGetWin32HandleInfoKHR

VkFenceGetWin32HandleInfoKHR - Structure describing a Win32 handle fence export operation

Description

The properties of the handle returned depend on the value of handleType. See VkExternalFenceHandleTypeFlagBits for a description of the properties of the defined external fence handle types.

Valid Usage

handleType must have been included in VkExportFenceCreateInfo::handleTypes when the fence’s current payload was created.

If handleType is defined as an NT handle, vkGetFenceWin32HandleKHR must be called no more than once for each valid unique combination of fence and handleType.

fence must not currently have its payload replaced by an imported payload as described below in Importing Fence Payloads unless that imported payload’s handle type was included in VkExternalFenceProperties::exportFromImportedHandleTypes for handleType.

If handleType refers to a handle type with copy payload transference semantics, fence must be signaled, or have an associated fence signal operation pending execution.

handleType must be defined as an NT handle or a global share handle.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_FENCE_GET_WIN32_HANDLE_INFO_KHR

pNext must be NULL

fence must be a valid VkFence handle

handleType must be a valid VkExternalFenceHandleTypeFlagBits value

See Also

VkExternalFenceHandleTypeFlagBits, VkFence, VkStructureType, vkGetFenceWin32HandleKHR

Constructors

VkFenceGetWin32HandleInfoKHR

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkFence :: VkFence

fence is the fence from which state will be exported.

vkHandleType :: VkExternalFenceHandleTypeFlagBits

handleType is the type of handle requested.

data VkImportFenceWin32HandleInfoKHR

VkImportFenceWin32HandleInfoKHR - (None)

Description

The handle types supported by handleType are:

+-----------------------+-----------------------+-----------------------+
| Handle Type           | Transference          | Permanence Supported  |
+=======================+=======================+=======================+
| 'Graphics.Vulkan.C.Co | Reference             | Temporary,Permanent   |
| re11.Promoted_from_VK |                       |                       |
| _KHR_external_fence_c |                       |                       |
| apabilities.VK_EXTERN |                       |                       |
| AL_FENCE_HANDLE_TYPE_ |                       |                       |
| OPAQUE_WIN32_BIT'     |                       |                       |
+-----------------------+-----------------------+-----------------------+
| 'Graphics.Vulkan.C.Co | Reference             | Temporary,Permanent   |
| re11.Promoted_from_VK |                       |                       |
| _KHR_external_fence_c |                       |                       |
| apabilities.VK_EXTERN |                       |                       |
| AL_FENCE_HANDLE_TYPE_ |                       |                       |
| OPAQUE_WIN32_KMT_BIT' |                       |                       |
+-----------------------+-----------------------+-----------------------+

Handle Types Supported by 'VkImportFenceWin32HandleInfoKHR'

Valid Usage

handleType must be a value included in the Handle Types Supported by VkImportFenceWin32HandleInfoKHR table.

If handleType is not VK_EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_BIT, name must be NULL.

If handleType is not 0 and handle is NULL, name must name a valid synchronization primitive of the type specified by handleType.

If handleType is not 0 and name is NULL, handle must be a valid handle of the type specified by handleType.

If handle is not NULL, name must be NULL.

If handle is not NULL, it must obey any requirements listed for handleType in external fence handle types compatibility.

If name is not NULL, it must obey any requirements listed for handleType in external fence handle types compatibility.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMPORT_FENCE_WIN32_HANDLE_INFO_KHR

pNext must be NULL

fence must be a valid VkFence handle

flags must be a valid combination of VkFenceImportFlagBits values

If handleType is not 0, handleType must be a valid VkExternalFenceHandleTypeFlagBits value

Host Synchronization

Host access to fence must be externally synchronized

See Also

VkExternalFenceHandleTypeFlagBits, VkFence, VkFenceImportFlags, VkStructureType, vkImportFenceWin32HandleKHR

Constructors

VkImportFenceWin32HandleInfoKHR

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkFence :: VkFence

fence is the fence into which the state will be imported.

vkFlags :: VkFenceImportFlags

flags is a bitmask of VkFenceImportFlagBits specifying additional parameters for the fence payload import operation.

vkHandleType :: VkExternalFenceHandleTypeFlagBits

handleType specifies the type of handle.

vkHandle :: HANDLE

handle is the external handle to import, or NULL.

vkName :: LPCWSTR

name is the NULL-terminated UTF-16 string naming the underlying synchronization primitive to import, or NULL.

type FN_vkGetFenceWin32HandleKHR = ("device" ::: VkDevice) -> ("pGetWin32HandleInfo" ::: Ptr VkFenceGetWin32HandleInfoKHR) -> ("pHandle" ::: Ptr HANDLE) -> IO VkResult

type PFN_vkGetFenceWin32HandleKHR = FunPtr FN_vkGetFenceWin32HandleKHR

vkGetFenceWin32HandleKHR :: DeviceCmds -> ("device" ::: VkDevice) -> ("pGetWin32HandleInfo" ::: Ptr VkFenceGetWin32HandleInfoKHR) -> ("pHandle" ::: Ptr HANDLE) -> IO VkResult

vkGetFenceWin32HandleKHR - Get a Windows HANDLE for a fence

Parameters

device is the logical device that created the fence being exported.

pGetWin32HandleInfo is a pointer to an instance of the VkFenceGetWin32HandleInfoKHR structure containing parameters of the export operation.

pHandle will return the Windows handle representing the fence state.

Description

For handle types defined as NT handles, the handles returned by vkGetFenceWin32HandleKHR are owned by the application. To avoid leaking resources, the application must release ownership of them using the CloseHandle system call when they are no longer needed.

Exporting a Windows handle from a fence may have side effects depending on the transference of the specified handle type, as described in Importing Fence Payloads.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_TOO_MANY_OBJECTS

VK_ERROR_OUT_OF_HOST_MEMORY

See Also

VkDevice, VkFenceGetWin32HandleInfoKHR

type FN_vkImportFenceWin32HandleKHR = ("device" ::: VkDevice) -> ("pImportFenceWin32HandleInfo" ::: Ptr VkImportFenceWin32HandleInfoKHR) -> IO VkResult

type PFN_vkImportFenceWin32HandleKHR = FunPtr FN_vkImportFenceWin32HandleKHR

vkImportFenceWin32HandleKHR :: DeviceCmds -> ("device" ::: VkDevice) -> ("pImportFenceWin32HandleInfo" ::: Ptr VkImportFenceWin32HandleInfoKHR) -> IO VkResult

vkImportFenceWin32HandleKHR - Import a fence from a Windows HANDLE

Parameters

device is the logical device that created the fence.

pImportFenceWin32HandleInfo points to a VkImportFenceWin32HandleInfoKHR structure specifying the fence and import parameters.

Description

Importing a fence payload from Windows handles does not transfer ownership of the handle to the Vulkan implementation. For handle types defined as NT handles, the application must release ownership using the CloseHandle system call when the handle is no longer needed.

Applications can import the same fence payload into multiple instances of Vulkan, into the same instance from which it was exported, and multiple times into a given Vulkan instance.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_INVALID_EXTERNAL_HANDLE

See Also

VkDevice, VkImportFenceWin32HandleInfoKHR

pattern VK_KHR_EXTERNAL_FENCE_WIN32_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_EXTERNAL_FENCE_WIN32_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_EXPORT_FENCE_WIN32_HANDLE_INFO_KHR :: VkStructureType

pattern VK_STRUCTURE_TYPE_FENCE_GET_WIN32_HANDLE_INFO_KHR :: VkStructureType

pattern VK_STRUCTURE_TYPE_IMPORT_FENCE_WIN32_HANDLE_INFO_KHR :: VkStructureType