Graphics.Vulkan.C.Extensions.VK_KHR_external_memory_fd

Documentation

data VkImportMemoryFdInfoKHR

VkImportMemoryFdInfoKHR - import memory created on the same physical device from a file descriptor

Description

Importing memory from a file descriptor transfers ownership of the file descriptor from the application to the Vulkan implementation. The application must not perform any operations on the file descriptor after a successful import.

Applications can import the same underlying memory into multiple instances of Vulkan, into the same instance from which it was exported, and multiple times into a given Vulkan instance. In all cases, each import operation must create a distinct VkDeviceMemory object.

Valid Usage

If handleType is not 0, it must be supported for import, as reported by VkExternalImageFormatProperties or VkExternalBufferProperties.

The memory from which fd was exported must have been created on the same underlying physical device as device.

If handleType is not 0, it must be defined as a POSIX file descriptor handle.

If handleType is not 0, fd must be a valid handle of the type specified by handleType.

The memory represented by fd must have been created from a physical device and driver that is compatible with device and handleType, as described in https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#external-memory-handle-types-compatibility.

fd must obey any requirements listed for handleType in external memory handle types compatibility.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMPORT_MEMORY_FD_INFO_KHR

If handleType is not 0, handleType must be a valid VkExternalMemoryHandleTypeFlagBits value

See Also

VkExternalMemoryHandleTypeFlagBits, VkStructureType

Constructors

VkImportMemoryFdInfoKHR

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkHandleType :: VkExternalMemoryHandleTypeFlagBits

handleType specifies the handle type of fd.

vkFd :: CInt

fd is the external handle to import.

data VkMemoryFdPropertiesKHR

VkMemoryFdPropertiesKHR - Properties of External Memory File Descriptors

Valid Usage (Implicit)

See Also

VkStructureType, vkGetMemoryFdPropertiesKHR

Constructors

VkMemoryFdPropertiesKHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_MEMORY_FD_PROPERTIES_KHR

vkPNext :: Ptr ()

pNext must be NULL

vkMemoryTypeBits :: Word32

memoryTypeBits is a bitmask containing one bit set for every memory type which the specified file descriptor can be imported as.

data VkMemoryGetFdInfoKHR

VkMemoryGetFdInfoKHR - Structure describing a POSIX FD semaphore export operation

Description

The properties of the file descriptor exported depend on the value of handleType. See VkExternalMemoryHandleTypeFlagBits for a description of the properties of the defined external memory handle types.

Valid Usage (Implicit)

See Also

VkDeviceMemory, VkExternalMemoryHandleTypeFlagBits, VkStructureType, vkGetMemoryFdKHR

Constructors

VkMemoryGetFdInfoKHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_MEMORY_GET_FD_INFO_KHR

vkPNext :: Ptr ()

pNext must be NULL

vkMemory :: VkDeviceMemory

memory must be a valid VkDeviceMemory handle

vkHandleType :: VkExternalMemoryHandleTypeFlagBits

handleType must be a valid VkExternalMemoryHandleTypeFlagBits value

type FN_vkGetMemoryFdKHR = ("device" ::: VkDevice) -> ("pGetFdInfo" ::: Ptr VkMemoryGetFdInfoKHR) -> ("pFd" ::: Ptr CInt) -> IO VkResult

type PFN_vkGetMemoryFdKHR = FunPtr FN_vkGetMemoryFdKHR

vkGetMemoryFdKHR :: DeviceCmds -> ("device" ::: VkDevice) -> ("pGetFdInfo" ::: Ptr VkMemoryGetFdInfoKHR) -> ("pFd" ::: Ptr CInt) -> IO VkResult

vkGetMemoryFdKHR - Get a POSIX file descriptor for a memory object

Parameters

device is the logical device that created the device memory being exported.

pGetFdInfo is a pointer to an instance of the VkMemoryGetFdInfoKHR structure containing parameters of the export operation.

pFd will return a file descriptor representing the underlying resources of the device memory object.

Description

Each call to vkGetMemoryFdKHR must create a new file descriptor and transfer ownership of it to the application. To avoid leaking resources, the application must release ownership of the file descriptor using the close system call when it is no longer needed, or by importing a Vulkan memory object from it. Where supported by the operating system, the implementation must set the file descriptor to be closed automatically when an execve system call is made.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_TOO_MANY_OBJECTS

VK_ERROR_OUT_OF_HOST_MEMORY

See Also

VkDevice, VkMemoryGetFdInfoKHR

type FN_vkGetMemoryFdPropertiesKHR = ("device" ::: VkDevice) -> ("handleType" ::: VkExternalMemoryHandleTypeFlagBits) -> ("fd" ::: CInt) -> ("pMemoryFdProperties" ::: Ptr VkMemoryFdPropertiesKHR) -> IO VkResult

type PFN_vkGetMemoryFdPropertiesKHR = FunPtr FN_vkGetMemoryFdPropertiesKHR

vkGetMemoryFdPropertiesKHR :: DeviceCmds -> ("device" ::: VkDevice) -> ("handleType" ::: VkExternalMemoryHandleTypeFlagBits) -> ("fd" ::: CInt) -> ("pMemoryFdProperties" ::: Ptr VkMemoryFdPropertiesKHR) -> IO VkResult

vkGetMemoryFdPropertiesKHR - Get Properties of External Memory File Descriptors

Parameters

device is the logical device that will be importing fd.

handleType is the type of the handle fd.

fd is the handle which will be imported.

pMemoryFdProperties is a pointer to a VkMemoryFdPropertiesKHR structure in which the properties of the handle fd are returned.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_INVALID_EXTERNAL_HANDLE

See Also

VkDevice, VkExternalMemoryHandleTypeFlagBits, VkMemoryFdPropertiesKHR

pattern VK_KHR_EXTERNAL_MEMORY_FD_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_EXTERNAL_MEMORY_FD_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_IMPORT_MEMORY_FD_INFO_KHR :: VkStructureType

pattern VK_STRUCTURE_TYPE_MEMORY_FD_PROPERTIES_KHR :: VkStructureType

pattern VK_STRUCTURE_TYPE_MEMORY_GET_FD_INFO_KHR :: VkStructureType