Graphics.Vulkan.C.Extensions.VK_KHR_external_semaphore

Documentation

type VkExportSemaphoreCreateInfoKHR = VkExportSemaphoreCreateInfo

pattern VkExportSemaphoreCreateInfoKHR :: ("sType" ::: VkStructureType) -> ("pNext" ::: Ptr ()) -> ("handleTypes" ::: VkExternalSemaphoreHandleTypeFlags) -> VkExportSemaphoreCreateInfoKHR

type VkSemaphoreImportFlagBitsKHR = VkSemaphoreImportFlagBits

type VkSemaphoreImportFlagsKHR = VkSemaphoreImportFlags

pattern VK_KHR_EXTERNAL_SEMAPHORE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_EXTERNAL_SEMAPHORE_SPEC_VERSION :: Integral a => a

pattern VK_SEMAPHORE_IMPORT_TEMPORARY_BIT_KHR :: VkSemaphoreImportFlagBits

pattern VK_STRUCTURE_TYPE_EXPORT_SEMAPHORE_CREATE_INFO_KHR :: VkStructureType

pattern VK_STRUCTURE_TYPE_EXPORT_SEMAPHORE_CREATE_INFO :: VkStructureType

pattern VK_SEMAPHORE_IMPORT_TEMPORARY_BIT :: VkSemaphoreImportFlagBits

VK_SEMAPHORE_IMPORT_TEMPORARY_BIT specifies that the semaphore payload will be imported only temporarily, as described in Importing Semaphore Payloads, regardless of the permanence of handleType.