Graphics.Vulkan.C.Extensions.VK_KHR_get_display_properties2

Documentation

data VkDisplayModeProperties2KHR

VkDisplayModeProperties2KHR - Structure describing an available display mode

Valid Usage (Implicit)

See Also

VkDisplayModePropertiesKHR, VkStructureType, vkGetDisplayModeProperties2KHR

Constructors

VkDisplayModeProperties2KHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DISPLAY_MODE_PROPERTIES_2_KHR

vkPNext :: Ptr ()

pNext must be NULL

vkDisplayModeProperties :: VkDisplayModePropertiesKHR

displayModeProperties is an instance of the VkDisplayModePropertiesKHR structure.

data VkDisplayPlaneCapabilities2KHR

VkDisplayPlaneCapabilities2KHR - Structure describing the capabilities of a mode and plane combination

Valid Usage (Implicit)

See Also

VkDisplayPlaneCapabilitiesKHR, VkStructureType, vkGetDisplayPlaneCapabilities2KHR

Constructors

VkDisplayPlaneCapabilities2KHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DISPLAY_PLANE_CAPABILITIES_2_KHR

vkPNext :: Ptr ()

pNext must be NULL

vkCapabilities :: VkDisplayPlaneCapabilitiesKHR

capabilities is an instance of the VkDisplayPlaneCapabilitiesKHR structure.

data VkDisplayPlaneInfo2KHR

VkDisplayPlaneInfo2KHR - Structure defining the intended configuration of a display plane

Description

Note

This parameter also implicitly specifies a display.

planeIndex is the plane which the application intends to use with the display.

The members of VkDisplayPlaneInfo2KHR correspond to the arguments to vkGetDisplayPlaneCapabilitiesKHR, with sType and pNext added for extensibility.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DISPLAY_PLANE_INFO_2_KHR

pNext must be NULL

mode must be a valid VkDisplayModeKHR handle

Host Synchronization

Host access to mode must be externally synchronized

See Also

VkDisplayModeKHR, VkStructureType, vkGetDisplayPlaneCapabilities2KHR

Constructors

VkDisplayPlaneInfo2KHR

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkMode :: VkDisplayModeKHR

mode is the display mode the application intends to program when using the specified plane.

vkPlaneIndex :: Word32

data VkDisplayPlaneProperties2KHR

VkDisplayPlaneProperties2KHR - Structure describing an available display plane

Valid Usage (Implicit)

See Also

VkDisplayPlanePropertiesKHR, VkStructureType, vkGetPhysicalDeviceDisplayPlaneProperties2KHR

Constructors

VkDisplayPlaneProperties2KHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DISPLAY_PLANE_PROPERTIES_2_KHR

vkPNext :: Ptr ()

pNext must be NULL

vkDisplayPlaneProperties :: VkDisplayPlanePropertiesKHR

displayPlaneProperties is an instance of the VkDisplayPlanePropertiesKHR structure.

data VkDisplayProperties2KHR

VkDisplayProperties2KHR - Structure describing an available display device

Valid Usage (Implicit)

See Also

VkDisplayPropertiesKHR, VkStructureType, vkGetPhysicalDeviceDisplayProperties2KHR

Constructors

VkDisplayProperties2KHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DISPLAY_PROPERTIES_2_KHR

vkPNext :: Ptr ()

pNext must be NULL

vkDisplayProperties :: VkDisplayPropertiesKHR

displayProperties is an instance of the VkDisplayPropertiesKHR structure.

type FN_vkGetDisplayModeProperties2KHR = ("physicalDevice" ::: VkPhysicalDevice) -> ("display" ::: VkDisplayKHR) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkDisplayModeProperties2KHR) -> IO VkResult

type PFN_vkGetDisplayModeProperties2KHR = FunPtr FN_vkGetDisplayModeProperties2KHR

vkGetDisplayModeProperties2KHR :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("display" ::: VkDisplayKHR) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkDisplayModeProperties2KHR) -> IO VkResult

vkGetDisplayModeProperties2KHR - Query information about the available display modes.

Parameters

physicalDevice is the physical device associated with display.

display is the display to query.

pPropertyCount is a pointer to an integer related to the number of display modes available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayModeProperties2KHR structures.

Description

vkGetDisplayModeProperties2KHR behaves similarly to vkGetDisplayModePropertiesKHR, with the ability to return extended information via chained output structures.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

display must be a valid VkDisplayKHR handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayModeProperties2KHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayKHR, VkDisplayModeProperties2KHR, VkPhysicalDevice

type FN_vkGetDisplayPlaneCapabilities2KHR = ("physicalDevice" ::: VkPhysicalDevice) -> ("pDisplayPlaneInfo" ::: Ptr VkDisplayPlaneInfo2KHR) -> ("pCapabilities" ::: Ptr VkDisplayPlaneCapabilities2KHR) -> IO VkResult

type PFN_vkGetDisplayPlaneCapabilities2KHR = FunPtr FN_vkGetDisplayPlaneCapabilities2KHR

vkGetDisplayPlaneCapabilities2KHR :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pDisplayPlaneInfo" ::: Ptr VkDisplayPlaneInfo2KHR) -> ("pCapabilities" ::: Ptr VkDisplayPlaneCapabilities2KHR) -> IO VkResult

vkGetDisplayPlaneCapabilities2KHR - Query capabilities of a mode and plane combination

Parameters

physicalDevice is the physical device associated with pDisplayPlaneInfo.

pDisplayPlaneInfo is a pointer to an instance of the VkDisplayPlaneInfo2KHR structure describing the plane and mode.

pCapabilities is a pointer to a VkDisplayPlaneCapabilities2KHR structure in which the capabilities are returned.

Description

vkGetDisplayPlaneCapabilities2KHR behaves similarly to vkGetDisplayPlaneCapabilitiesKHR, with the ability to specify extended inputs via chained input structures, and to return extended information via chained output structures.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayPlaneCapabilities2KHR, VkDisplayPlaneInfo2KHR, VkPhysicalDevice

type FN_vkGetPhysicalDeviceDisplayPlaneProperties2KHR = ("physicalDevice" ::: VkPhysicalDevice) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkDisplayPlaneProperties2KHR) -> IO VkResult

type PFN_vkGetPhysicalDeviceDisplayPlaneProperties2KHR = FunPtr FN_vkGetPhysicalDeviceDisplayPlaneProperties2KHR

vkGetPhysicalDeviceDisplayPlaneProperties2KHR :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkDisplayPlaneProperties2KHR) -> IO VkResult

vkGetPhysicalDeviceDisplayPlaneProperties2KHR - Query information about the available display planes.

Parameters

physicalDevice is a physical device.

pPropertyCount is a pointer to an integer related to the number of display planes available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayPlaneProperties2KHR structures.

Description

vkGetPhysicalDeviceDisplayPlaneProperties2KHR behaves similarly to vkGetPhysicalDeviceDisplayPlanePropertiesKHR, with the ability to return extended information via chained output structures.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayPlaneProperties2KHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayPlaneProperties2KHR, VkPhysicalDevice

type FN_vkGetPhysicalDeviceDisplayProperties2KHR = ("physicalDevice" ::: VkPhysicalDevice) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkDisplayProperties2KHR) -> IO VkResult

type PFN_vkGetPhysicalDeviceDisplayProperties2KHR = FunPtr FN_vkGetPhysicalDeviceDisplayProperties2KHR

vkGetPhysicalDeviceDisplayProperties2KHR :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkDisplayProperties2KHR) -> IO VkResult

vkGetPhysicalDeviceDisplayProperties2KHR - Query information about the available displays

Parameters

physicalDevice is a physical device.

pPropertyCount is a pointer to an integer related to the number of display devices available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayProperties2KHR structures.

Description

vkGetPhysicalDeviceDisplayProperties2KHR behaves similarly to vkGetPhysicalDeviceDisplayPropertiesKHR, with the ability to return extended information via chained output structures.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayProperties2KHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayProperties2KHR, VkPhysicalDevice

pattern VK_KHR_GET_DISPLAY_PROPERTIES_2_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_GET_DISPLAY_PROPERTIES_2_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_DISPLAY_MODE_PROPERTIES_2_KHR :: VkStructureType

pattern VK_STRUCTURE_TYPE_DISPLAY_PLANE_CAPABILITIES_2_KHR :: VkStructureType

pattern VK_STRUCTURE_TYPE_DISPLAY_PLANE_INFO_2_KHR :: VkStructureType

pattern VK_STRUCTURE_TYPE_DISPLAY_PLANE_PROPERTIES_2_KHR :: VkStructureType

pattern VK_STRUCTURE_TYPE_DISPLAY_PROPERTIES_2_KHR :: VkStructureType