Graphics.Vulkan.C.Extensions.VK_KHR_get_surface_capabilities2

Documentation

data VkPhysicalDeviceSurfaceInfo2KHR

VkPhysicalDeviceSurfaceInfo2KHR - Structure specifying a surface and related swapchain creation parameters

Description

The members of VkPhysicalDeviceSurfaceInfo2KHR correspond to the arguments to vkGetPhysicalDeviceSurfaceCapabilitiesKHR, with sType and pNext added for extensibility.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SURFACE_INFO_2_KHR

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkSurfaceFullScreenExclusiveInfoEXT or VkSurfaceFullScreenExclusiveWin32InfoEXT

Each sType member in the pNext chain must be unique

surface must be a valid VkSurfaceKHR handle

See Also

VkStructureType, VkSurfaceKHR, vkGetDeviceGroupSurfacePresentModes2EXT, vkGetPhysicalDeviceSurfaceCapabilities2KHR, vkGetPhysicalDeviceSurfaceFormats2KHR, vkGetPhysicalDeviceSurfacePresentModes2EXT

Constructors

VkPhysicalDeviceSurfaceInfo2KHR

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkSurface :: VkSurfaceKHR

surface is the surface that will be associated with the swapchain.

data VkSurfaceCapabilities2KHR

VkSurfaceCapabilities2KHR - Structure describing capabilities of a surface

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_SURFACE_CAPABILITIES_2_KHR

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkDisplayNativeHdrSurfaceCapabilitiesAMD, VkSharedPresentSurfaceCapabilitiesKHR, VkSurfaceCapabilitiesFullScreenExclusiveEXT, or VkSurfaceProtectedCapabilitiesKHR

Each sType member in the pNext chain must be unique

See Also

VkStructureType, VkSurfaceCapabilitiesKHR, vkGetPhysicalDeviceSurfaceCapabilities2KHR

Constructors

VkSurfaceCapabilities2KHR

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkSurfaceCapabilities :: VkSurfaceCapabilitiesKHR

surfaceCapabilities is a structure of type VkSurfaceCapabilitiesKHR describing the capabilities of the specified surface.

data VkSurfaceFormat2KHR

VkSurfaceFormat2KHR - Structure describing a supported swapchain format tuple

Valid Usage (Implicit)

See Also

VkStructureType, VkSurfaceFormatKHR, vkGetPhysicalDeviceSurfaceFormats2KHR

Constructors

VkSurfaceFormat2KHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_SURFACE_FORMAT_2_KHR

vkPNext :: Ptr ()

pNext must be NULL

vkSurfaceFormat :: VkSurfaceFormatKHR

surfaceFormat is an instance of VkSurfaceFormatKHR describing a format-color space pair that is compatible with the specified surface.

type FN_vkGetPhysicalDeviceSurfaceCapabilities2KHR = ("physicalDevice" ::: VkPhysicalDevice) -> ("pSurfaceInfo" ::: Ptr VkPhysicalDeviceSurfaceInfo2KHR) -> ("pSurfaceCapabilities" ::: Ptr VkSurfaceCapabilities2KHR) -> IO VkResult

type PFN_vkGetPhysicalDeviceSurfaceCapabilities2KHR = FunPtr FN_vkGetPhysicalDeviceSurfaceCapabilities2KHR

vkGetPhysicalDeviceSurfaceCapabilities2KHR :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pSurfaceInfo" ::: Ptr VkPhysicalDeviceSurfaceInfo2KHR) -> ("pSurfaceCapabilities" ::: Ptr VkSurfaceCapabilities2KHR) -> IO VkResult

vkGetPhysicalDeviceSurfaceCapabilities2KHR - Reports capabilities of a surface on a physical device

Parameters

physicalDevice is the physical device that will be associated with the swapchain to be created, as described for vkCreateSwapchainKHR.

pSurfaceInfo points to an instance of the VkPhysicalDeviceSurfaceInfo2KHR structure, describing the surface and other fixed parameters that would be consumed by vkCreateSwapchainKHR.

pSurfaceCapabilities points to an instance of the VkSurfaceCapabilities2KHR structure in which the capabilities are returned.

Description

vkGetPhysicalDeviceSurfaceCapabilities2KHR behaves similarly to vkGetPhysicalDeviceSurfaceCapabilitiesKHR, with the ability to specify extended inputs via chained input structures, and to return extended information via chained output structures.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

See Also

VkPhysicalDevice, VkPhysicalDeviceSurfaceInfo2KHR, VkSurfaceCapabilities2KHR

type FN_vkGetPhysicalDeviceSurfaceFormats2KHR = ("physicalDevice" ::: VkPhysicalDevice) -> ("pSurfaceInfo" ::: Ptr VkPhysicalDeviceSurfaceInfo2KHR) -> ("pSurfaceFormatCount" ::: Ptr Word32) -> ("pSurfaceFormats" ::: Ptr VkSurfaceFormat2KHR) -> IO VkResult

type PFN_vkGetPhysicalDeviceSurfaceFormats2KHR = FunPtr FN_vkGetPhysicalDeviceSurfaceFormats2KHR

vkGetPhysicalDeviceSurfaceFormats2KHR :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pSurfaceInfo" ::: Ptr VkPhysicalDeviceSurfaceInfo2KHR) -> ("pSurfaceFormatCount" ::: Ptr Word32) -> ("pSurfaceFormats" ::: Ptr VkSurfaceFormat2KHR) -> IO VkResult

vkGetPhysicalDeviceSurfaceFormats2KHR - Query color formats supported by surface

Parameters

physicalDevice is the physical device that will be associated with the swapchain to be created, as described for vkCreateSwapchainKHR.

pSurfaceInfo points to an instance of the VkPhysicalDeviceSurfaceInfo2KHR structure, describing the surface and other fixed parameters that would be consumed by vkCreateSwapchainKHR.

pSurfaceFormatCount is a pointer to an integer related to the number of format tuples available or queried, as described below.

pSurfaceFormats is either NULL or a pointer to an array of VkSurfaceFormat2KHR structures.

Description

If pSurfaceFormats is NULL, then the number of format tuples supported for the given surface is returned in pSurfaceFormatCount. The number of format tuples supported will be greater than or equal to 1. Otherwise, pSurfaceFormatCount must point to a variable set by the user to the number of elements in the pSurfaceFormats array, and on return the variable is overwritten with the number of structures actually written to pSurfaceFormats. If the value of pSurfaceFormatCount is less than the number of format tuples supported, at most pSurfaceFormatCount structures will be written. If pSurfaceFormatCount is smaller than the number of format tuples supported for the surface parameters described in pSurfaceInfo, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pSurfaceInfo must be a valid pointer to a valid VkPhysicalDeviceSurfaceInfo2KHR structure

pSurfaceFormatCount must be a valid pointer to a uint32_t value

If the value referenced by pSurfaceFormatCount is not 0, and pSurfaceFormats is not NULL, pSurfaceFormats must be a valid pointer to an array of pSurfaceFormatCount VkSurfaceFormat2KHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

See Also

VkPhysicalDevice, VkPhysicalDeviceSurfaceInfo2KHR, VkSurfaceFormat2KHR

pattern VK_KHR_GET_SURFACE_CAPABILITIES_2_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_GET_SURFACE_CAPABILITIES_2_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SURFACE_INFO_2_KHR :: VkStructureType

pattern VK_STRUCTURE_TYPE_SURFACE_CAPABILITIES_2_KHR :: VkStructureType

pattern VK_STRUCTURE_TYPE_SURFACE_FORMAT_2_KHR :: VkStructureType