Graphics.Vulkan.C.Extensions.VK_KHR_incremental_present

Documentation

data VkPresentRegionKHR

VkPresentRegionKHR - Structure containing rectangular region changed by vkQueuePresentKHR for a given VkImage

Valid Usage (Implicit)

If rectangleCount is not 0, and pRectangles is not NULL, pRectangles must be a valid pointer to an array of rectangleCount valid VkRectLayerKHR structures

See Also

VkPresentRegionsKHR, VkRectLayerKHR

Constructors

VkPresentRegionKHR

Fields

vkRectangleCount :: Word32

rectangleCount is the number of rectangles in pRectangles, or zero if the entire image has changed and should be presented.

vkPRectangles :: Ptr VkRectLayerKHR

pRectangles is either NULL or a pointer to an array of VkRectLayerKHR structures. The VkRectLayerKHR structure is the framebuffer coordinates, plus layer, of a portion of a presentable image that has changed and must be presented. If non-NULL, each entry in pRectangles is a rectangle of the given image that has changed since the last image was presented to the given swapchain.

data VkPresentRegionsKHR

VkPresentRegionsKHR - Structure hint of rectangular regions changed by vkQueuePresentKHR

Valid Usage

swapchainCount must be the same value as VkPresentInfoKHR::swapchainCount, where VkPresentInfoKHR is in the pNext chain of this VkPresentRegionsKHR structure

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PRESENT_REGIONS_KHR

If pRegions is not NULL, pRegions must be a valid pointer to an array of swapchainCount valid VkPresentRegionKHR structures

swapchainCount must be greater than 0

See Also

VkPresentRegionKHR, VkStructureType

Constructors

VkPresentRegionsKHR

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkSwapchainCount :: Word32

swapchainCount is the number of swapchains being presented to by this command.

vkPRegions :: Ptr VkPresentRegionKHR

pRegions is NULL or a pointer to an array of VkPresentRegionKHR elements with swapchainCount entries. If not NULL, each element of pRegions contains the region that has changed since the last present to the swapchain in the corresponding entry in the VkPresentInfoKHR::pSwapchains array.

data VkRectLayerKHR

VkRectLayerKHR - Structure containing a rectangle, including layer, changed by vkQueuePresentKHR for a given VkImage

Valid Usage

The sum of offset and extent must be no greater than the imageExtent member of the VkSwapchainCreateInfoKHR structure given to vkCreateSwapchainKHR.

layer must be less than imageArrayLayers member of the VkSwapchainCreateInfoKHR structure given to vkCreateSwapchainKHR.

Some platforms allow the size of a surface to change, and then scale the pixels of the image to fit the surface. VkRectLayerKHR specifies pixels of the swapchain’s image(s), which will be constant for the life of the swapchain.

See Also

VkExtent2D, VkOffset2D, VkPresentRegionKHR

Constructors

VkRectLayerKHR

Fields

vkOffset :: VkOffset2D

offset is the origin of the rectangle, in pixels.

vkExtent :: VkExtent2D

extent is the size of the rectangle, in pixels.

vkLayer :: Word32

layer is the layer of the image. For images with only one layer, the value of layer must be 0.

pattern VK_KHR_INCREMENTAL_PRESENT_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_INCREMENTAL_PRESENT_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_PRESENT_REGIONS_KHR :: VkStructureType