Graphics.Vulkan.C.Extensions.VK_KHR_shader_atomic_int64

Documentation

data VkPhysicalDeviceShaderAtomicInt64FeaturesKHR

VkPhysicalDeviceShaderAtomicInt64FeaturesKHR - Structure describing features supported by VK_KHR_shader_atomic_int64

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

VkPhysicalDeviceShaderAtomicInt64FeaturesKHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_ATOMIC_INT64_FEATURES_KHR

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkShaderBufferInt64Atomics :: VkBool32

shaderBufferInt64Atomics indicates whether shaders can support 64-bit unsigned and signed integer atomic operations on buffers.

vkShaderSharedInt64Atomics :: VkBool32

shaderSharedInt64Atomics indicates whether shaders can support 64-bit unsigned and signed integer atomic operations on shared memory.

pattern VK_KHR_SHADER_ATOMIC_INT64_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_SHADER_ATOMIC_INT64_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_ATOMIC_INT64_FEATURES_KHR :: VkStructureType