Graphics.Vulkan.C.Extensions.VK_KHR_shared_presentable_image

Documentation

data VkSharedPresentSurfaceCapabilitiesKHR

VkSharedPresentSurfaceCapabilitiesKHR - structure describing capabilities of a surface for shared presentation

Valid Usage (Implicit)

See Also

VkImageUsageFlags, VkStructureType

Constructors

VkSharedPresentSurfaceCapabilitiesKHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_SHARED_PRESENT_SURFACE_CAPABILITIES_KHR

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkSharedPresentSupportedUsageFlags :: VkImageUsageFlags

sharedPresentSupportedUsageFlags is a bitmask of VkImageUsageFlagBits representing the ways the application can use the shared presentable image from a swapchain created with VkPresentModeKHR set to VK_PRESENT_MODE_SHARED_DEMAND_REFRESH_KHR or VK_PRESENT_MODE_SHARED_CONTINUOUS_REFRESH_KHR for the surface on the specified device. VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT must be included in the set but implementations may support additional usages.

type FN_vkGetSwapchainStatusKHR = ("device" ::: VkDevice) -> ("swapchain" ::: VkSwapchainKHR) -> IO VkResult

type PFN_vkGetSwapchainStatusKHR = FunPtr FN_vkGetSwapchainStatusKHR

vkGetSwapchainStatusKHR :: DeviceCmds -> ("device" ::: VkDevice) -> ("swapchain" ::: VkSwapchainKHR) -> IO VkResult

vkGetSwapchainStatusKHR - Get a swapchain’s status

Parameters

device is the device associated with swapchain.

swapchain is the swapchain to query.

Valid Usage (Implicit)

device must be a valid VkDevice handle

swapchain must be a valid VkSwapchainKHR handle

Both of device, and swapchain must have been created, allocated, or retrieved from the same VkInstance

Host Synchronization

Host access to swapchain must be externally synchronized

Return Codes

Success

- VK_SUCCESS

VK_SUBOPTIMAL_KHR

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_DEVICE_LOST

VK_ERROR_OUT_OF_DATE_KHR

VK_ERROR_SURFACE_LOST_KHR

VK_ERROR_FULL_SCREEN_EXCLUSIVE_MODE_LOST_EXT

See Also

VkDevice, VkSwapchainKHR

pattern VK_IMAGE_LAYOUT_SHARED_PRESENT_KHR :: VkImageLayout

pattern VK_KHR_SHARED_PRESENTABLE_IMAGE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_SHARED_PRESENTABLE_IMAGE_SPEC_VERSION :: Integral a => a

pattern VK_PRESENT_MODE_SHARED_CONTINUOUS_REFRESH_KHR :: VkPresentModeKHR

pattern VK_PRESENT_MODE_SHARED_DEMAND_REFRESH_KHR :: VkPresentModeKHR

pattern VK_STRUCTURE_TYPE_SHARED_PRESENT_SURFACE_CAPABILITIES_KHR :: VkStructureType