Graphics.Vulkan.C.Extensions.VK_KHR_vulkan_memory_model

Documentation

data VkPhysicalDeviceVulkanMemoryModelFeaturesKHR

VkPhysicalDeviceVulkanMemoryModelFeaturesKHR - Structure describing features supported by VK_KHR_vulkan_memory_model

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

VkPhysicalDeviceVulkanMemoryModelFeaturesKHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VULKAN_MEMORY_MODEL_FEATURES_KHR

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkVulkanMemoryModel :: VkBool32

vulkanMemoryModel indicates whether the Vulkan Memory Model is supported, as defined in Vulkan Memory Model. This also indicates whether shader modules can declare the VulkanMemoryModelKHR capability.

vkVulkanMemoryModelDeviceScope :: VkBool32

vulkanMemoryModelDeviceScope indicates whether the Vulkan Memory Model can use Device scope synchronization. This also indicates whether shader modules can declare the VulkanMemoryModelDeviceScopeKHR capability.

vkVulkanMemoryModelAvailabilityVisibilityChains :: VkBool32

vulkanMemoryModelAvailabilityVisibilityChains indicates whether the Vulkan Memory Model can use availability and visibility chains with more than one element.

pattern VK_KHR_VULKAN_MEMORY_MODEL_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_VULKAN_MEMORY_MODEL_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VULKAN_MEMORY_MODEL_FEATURES_KHR :: VkStructureType