Graphics.Vulkan.C.Extensions.VK_KHR_wayland_surface

Documentation

newtype VkWaylandSurfaceCreateFlagsKHR

Constructors

VkWaylandSurfaceCreateFlagsKHR VkFlags

data VkWaylandSurfaceCreateInfoKHR

VkWaylandSurfaceCreateInfoKHR - Structure specifying parameters of a newly created Wayland surface object

Valid Usage (Implicit)

See Also

VkStructureType, VkWaylandSurfaceCreateFlagsKHR, vkCreateWaylandSurfaceKHR

Constructors

VkWaylandSurfaceCreateInfoKHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_WAYLAND_SURFACE_CREATE_INFO_KHR

vkPNext :: Ptr ()

pNext must be NULL

vkFlags :: VkWaylandSurfaceCreateFlagsKHR

flags must be 0

vkDisplay :: Ptr Wl_display

display must point to a valid Wayland wl_display.

vkSurface :: Ptr Wl_surface

surface must point to a valid Wayland wl_surface.

data Wl_display

Opaque data

data Wl_surface

Opaque data

type FN_vkCreateWaylandSurfaceKHR = ("instance" ::: VkInstance) -> ("pCreateInfo" ::: Ptr VkWaylandSurfaceCreateInfoKHR) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pSurface" ::: Ptr VkSurfaceKHR) -> IO VkResult

type PFN_vkCreateWaylandSurfaceKHR = FunPtr FN_vkCreateWaylandSurfaceKHR

vkCreateWaylandSurfaceKHR :: InstanceCmds -> ("instance" ::: VkInstance) -> ("pCreateInfo" ::: Ptr VkWaylandSurfaceCreateInfoKHR) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pSurface" ::: Ptr VkSurfaceKHR) -> IO VkResult

vkCreateWaylandSurfaceKHR - Create a VkSurfaceKHR object for a Wayland window

Parameters

instance is the instance to associate the surface with.

pCreateInfo is a pointer to an instance of the VkWaylandSurfaceCreateInfoKHR structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkWaylandSurfaceCreateInfoKHR structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkInstance, VkSurfaceKHR, VkWaylandSurfaceCreateInfoKHR

type FN_vkGetPhysicalDeviceWaylandPresentationSupportKHR = ("physicalDevice" ::: VkPhysicalDevice) -> ("queueFamilyIndex" ::: Word32) -> ("display" ::: Ptr Wl_display) -> IO VkBool32

type PFN_vkGetPhysicalDeviceWaylandPresentationSupportKHR = FunPtr FN_vkGetPhysicalDeviceWaylandPresentationSupportKHR

vkGetPhysicalDeviceWaylandPresentationSupportKHR :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("queueFamilyIndex" ::: Word32) -> ("display" ::: Ptr Wl_display) -> IO VkBool32

vkGetPhysicalDeviceWaylandPresentationSupportKHR - Query physical device for presentation to Wayland

Parameters

physicalDevice is the physical device.

queueFamilyIndex is the queue family index.

display is a pointer to the wl_display associated with a Wayland compositor.

Description

This platform-specific function can be called prior to creating a surface.

Valid Usage (Implicit)

See Also

VkPhysicalDevice

pattern VK_KHR_WAYLAND_SURFACE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_WAYLAND_SURFACE_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_WAYLAND_SURFACE_CREATE_INFO_KHR :: VkStructureType