Graphics.Vulkan.C.Extensions.VK_KHR_win32_surface

Documentation

type HINSTANCE = Ptr ()

type HWND = Ptr ()

newtype VkWin32SurfaceCreateFlagsKHR

Constructors

VkWin32SurfaceCreateFlagsKHR VkFlags

data VkWin32SurfaceCreateInfoKHR

VkWin32SurfaceCreateInfoKHR - Structure specifying parameters of a newly created Win32 surface object

Valid Usage (Implicit)

See Also

VkStructureType, VkWin32SurfaceCreateFlagsKHR, vkCreateWin32SurfaceKHR

Constructors

VkWin32SurfaceCreateInfoKHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_WIN32_SURFACE_CREATE_INFO_KHR

vkPNext :: Ptr ()

pNext must be NULL

vkFlags :: VkWin32SurfaceCreateFlagsKHR

flags must be 0

vkHinstance :: HINSTANCE

hinstance must be a valid Win32 HINSTANCE.

vkHwnd :: HWND

hwnd must be a valid Win32 HWND.

type FN_vkCreateWin32SurfaceKHR = ("instance" ::: VkInstance) -> ("pCreateInfo" ::: Ptr VkWin32SurfaceCreateInfoKHR) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pSurface" ::: Ptr VkSurfaceKHR) -> IO VkResult

type PFN_vkCreateWin32SurfaceKHR = FunPtr FN_vkCreateWin32SurfaceKHR

vkCreateWin32SurfaceKHR :: InstanceCmds -> ("instance" ::: VkInstance) -> ("pCreateInfo" ::: Ptr VkWin32SurfaceCreateInfoKHR) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pSurface" ::: Ptr VkSurfaceKHR) -> IO VkResult

vkCreateWin32SurfaceKHR - Create a VkSurfaceKHR object for an Win32 native window

Parameters

instance is the instance to associate the surface with.

pCreateInfo is a pointer to an instance of the VkWin32SurfaceCreateInfoKHR structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkWin32SurfaceCreateInfoKHR structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkInstance, VkSurfaceKHR, VkWin32SurfaceCreateInfoKHR

type FN_vkGetPhysicalDeviceWin32PresentationSupportKHR = ("physicalDevice" ::: VkPhysicalDevice) -> ("queueFamilyIndex" ::: Word32) -> IO VkBool32

type PFN_vkGetPhysicalDeviceWin32PresentationSupportKHR = FunPtr FN_vkGetPhysicalDeviceWin32PresentationSupportKHR

vkGetPhysicalDeviceWin32PresentationSupportKHR :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("queueFamilyIndex" ::: Word32) -> IO VkBool32

vkGetPhysicalDeviceWin32PresentationSupportKHR - query queue family support for presentation on a Win32 display

Parameters

physicalDevice is the physical device.

queueFamilyIndex is the queue family index.

Description

This platform-specific function can be called prior to creating a surface.

Valid Usage (Implicit)

See Also

VkPhysicalDevice

pattern VK_KHR_WIN32_SURFACE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_WIN32_SURFACE_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_WIN32_SURFACE_CREATE_INFO_KHR :: VkStructureType