Graphics.Vulkan.C.Extensions.VK_KHR_xcb_surface

Documentation

newtype VkXcbSurfaceCreateFlagsKHR

Constructors

VkXcbSurfaceCreateFlagsKHR VkFlags

data VkXcbSurfaceCreateInfoKHR

VkXcbSurfaceCreateInfoKHR - Structure specifying parameters of a newly created Xcb surface object

Valid Usage (Implicit)

See Also

VkStructureType, VkXcbSurfaceCreateFlagsKHR, vkCreateXcbSurfaceKHR

Constructors

VkXcbSurfaceCreateInfoKHR

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_XCB_SURFACE_CREATE_INFO_KHR

vkPNext :: Ptr ()

pNext must be NULL

vkFlags :: VkXcbSurfaceCreateFlagsKHR

flags must be 0

vkConnection :: Ptr Xcb_connection_t

connection must point to a valid X11 xcb_connection_t.

vkWindow :: Xcb_window_t

window must be a valid X11 xcb_window_t.

data Xcb_connection_t

Opaque data

type Xcb_visualid_t = Word32

type Xcb_window_t = Word32

type FN_vkCreateXcbSurfaceKHR = ("instance" ::: VkInstance) -> ("pCreateInfo" ::: Ptr VkXcbSurfaceCreateInfoKHR) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pSurface" ::: Ptr VkSurfaceKHR) -> IO VkResult

type PFN_vkCreateXcbSurfaceKHR = FunPtr FN_vkCreateXcbSurfaceKHR

vkCreateXcbSurfaceKHR :: InstanceCmds -> ("instance" ::: VkInstance) -> ("pCreateInfo" ::: Ptr VkXcbSurfaceCreateInfoKHR) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pSurface" ::: Ptr VkSurfaceKHR) -> IO VkResult

vkCreateXcbSurfaceKHR - Create a VkSurfaceKHR object for a X11 window, using the XCB client-side library

Parameters

instance is the instance to associate the surface with.

pCreateInfo is a pointer to an instance of the VkXcbSurfaceCreateInfoKHR structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkXcbSurfaceCreateInfoKHR structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkInstance, VkSurfaceKHR, VkXcbSurfaceCreateInfoKHR

type FN_vkGetPhysicalDeviceXcbPresentationSupportKHR = ("physicalDevice" ::: VkPhysicalDevice) -> ("queueFamilyIndex" ::: Word32) -> ("connection" ::: Ptr Xcb_connection_t) -> ("visual_id" ::: Xcb_visualid_t) -> IO VkBool32

type PFN_vkGetPhysicalDeviceXcbPresentationSupportKHR = FunPtr FN_vkGetPhysicalDeviceXcbPresentationSupportKHR

vkGetPhysicalDeviceXcbPresentationSupportKHR :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("queueFamilyIndex" ::: Word32) -> ("connection" ::: Ptr Xcb_connection_t) -> ("visual_id" ::: Xcb_visualid_t) -> IO VkBool32

vkGetPhysicalDeviceXcbPresentationSupportKHR - Query physical device for presentation to X11 server using XCB

Parameters

physicalDevice is the physical device.

queueFamilyIndex is the queue family index.

connection is a pointer to an xcb_connection_t to the X server. visual_id is an X11 visual (xcb_visualid_t).

Description

This platform-specific function can be called prior to creating a surface.

Valid Usage (Implicit)

See Also

VkPhysicalDevice

pattern VK_KHR_XCB_SURFACE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_KHR_XCB_SURFACE_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_XCB_SURFACE_CREATE_INFO_KHR :: VkStructureType