Graphics.Vulkan.C.Extensions.VK_MVK_ios_surface

Documentation

newtype VkIOSSurfaceCreateFlagsMVK

Constructors

VkIOSSurfaceCreateFlagsMVK VkFlags

data VkIOSSurfaceCreateInfoMVK

VkIOSSurfaceCreateInfoMVK - Structure specifying parameters of a newly created iOS surface object

Valid Usage (Implicit)

See Also

VkIOSSurfaceCreateFlagsMVK, VkStructureType, vkCreateIOSSurfaceMVK

Constructors

VkIOSSurfaceCreateInfoMVK

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_IOS_SURFACE_CREATE_INFO_MVK

vkPNext :: Ptr ()

pNext must be NULL

vkFlags :: VkIOSSurfaceCreateFlagsMVK

flags must be 0

vkPView :: Ptr ()

pView must be a valid UIView and must be backed by a CALayer instance of type CAMetalLayer.

type FN_vkCreateIOSSurfaceMVK = ("instance" ::: VkInstance) -> ("pCreateInfo" ::: Ptr VkIOSSurfaceCreateInfoMVK) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pSurface" ::: Ptr VkSurfaceKHR) -> IO VkResult

type PFN_vkCreateIOSSurfaceMVK = FunPtr FN_vkCreateIOSSurfaceMVK

vkCreateIOSSurfaceMVK :: InstanceCmds -> ("instance" ::: VkInstance) -> ("pCreateInfo" ::: Ptr VkIOSSurfaceCreateInfoMVK) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pSurface" ::: Ptr VkSurfaceKHR) -> IO VkResult

vkCreateIOSSurfaceMVK - Create a VkSurfaceKHR object for an iOS UIView

Parameters

instance is the instance with which to associate the surface.

pCreateInfo is a pointer to an instance of the VkIOSSurfaceCreateInfoMVK structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkIOSSurfaceCreateInfoMVK structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_NATIVE_WINDOW_IN_USE_KHR

See Also

VkAllocationCallbacks, VkIOSSurfaceCreateInfoMVK, VkInstance, VkSurfaceKHR

pattern VK_MVK_IOS_SURFACE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_MVK_IOS_SURFACE_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_IOS_SURFACE_CREATE_INFO_MVK :: VkStructureType