Graphics.Vulkan.C.Extensions.VK_NVX_device_generated_commands

Documentation

data VkCmdProcessCommandsInfoNVX

VkCmdProcessCommandsInfoNVX - Structure specifying parameters for the generation of commands

Valid Usage

The provided objectTable must include all objects referenced by the generation process

indirectCommandsTokenCount must match the indirectCommandsLayout’s tokenCount

The tokenType member of each entry in the pIndirectCommandsTokens array must match the values used at creation time of indirectCommandsLayout

If targetCommandBuffer is provided, it must have reserved command space

If targetCommandBuffer is provided, the objectTable must match the reservation’s objectTable and must have had all referenced objects registered at reservation time

If targetCommandBuffer is provided, the indirectCommandsLayout must match the reservation’s indirectCommandsLayout

If targetCommandBuffer is provided, the maxSequencesCount must not exceed the reservation’s maxSequencesCount

If sequencesCountBuffer is used, its usage flag must have the VK_BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set

If sequencesCountBuffer is used, sequencesCountOffset must be aligned to VkDeviceGeneratedCommandsLimitsNVX::minSequenceCountBufferOffsetAlignment

If sequencesIndexBuffer is used, its usage flag must have the VK_BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set

If sequencesIndexBuffer is used, sequencesIndexOffset must be aligned to VkDeviceGeneratedCommandsLimitsNVX::minSequenceIndexBufferOffsetAlignment

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_CMD_PROCESS_COMMANDS_INFO_NVX

pNext must be NULL

objectTable must be a valid VkObjectTableNVX handle

indirectCommandsLayout must be a valid VkIndirectCommandsLayoutNVX handle

pIndirectCommandsTokens must be a valid pointer to an array of indirectCommandsTokenCount valid VkIndirectCommandsTokenNVX structures

If targetCommandBuffer is not NULL, targetCommandBuffer must be a valid VkCommandBuffer handle

If sequencesCountBuffer is not VK_NULL_HANDLE, sequencesCountBuffer must be a valid VkBuffer handle

If sequencesIndexBuffer is not VK_NULL_HANDLE, sequencesIndexBuffer must be a valid VkBuffer handle

indirectCommandsTokenCount must be greater than 0

Each of indirectCommandsLayout, objectTable, sequencesCountBuffer, sequencesIndexBuffer, and targetCommandBuffer that are valid handles must have been created, allocated, or retrieved from the same VkDevice

Host Synchronization

Host access to objectTable must be externally synchronized

Host access to targetCommandBuffer must be externally synchronized

See Also

VkBuffer, VkCommandBuffer, VkDeviceSize, VkIndirectCommandsLayoutNVX, VkIndirectCommandsTokenNVX, VkObjectTableNVX, VkStructureType, vkCmdProcessCommandsNVX

Constructors

VkCmdProcessCommandsInfoNVX

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkObjectTable :: VkObjectTableNVX

objectTable is the VkObjectTableNVX to be used for the generation process. Only registered objects at the time vkCmdReserveSpaceForCommandsNVX is called, will be taken into account for the reservation.

vkIndirectCommandsLayout :: VkIndirectCommandsLayoutNVX

indirectCommandsLayout is the VkIndirectCommandsLayoutNVX that provides the command sequence to generate.

vkIndirectCommandsTokenCount :: Word32

indirectCommandsTokenCount defines the number of input tokens used.

vkPIndirectCommandsTokens :: Ptr VkIndirectCommandsTokenNVX

pIndirectCommandsTokens provides an array of VkIndirectCommandsTokenNVX that reference the input data for each token command.

vkMaxSequencesCount :: Word32

maxSequencesCount is the maximum number of sequences for which command buffer space will be reserved. If sequencesCountBuffer is VK_NULL_HANDLE, this is also the actual number of sequences generated.

vkTargetCommandBuffer :: VkCommandBuffer

targetCommandBuffer can be the secondary VkCommandBuffer in which the commands should be recorded. If targetCommandBuffer is NULL an implicit reservation as well as execution takes place on the processing VkCommandBuffer.

vkSequencesCountBuffer :: VkBuffer

sequencesCountBuffer can be VkBuffer from which the actual amount of sequences is sourced from as uint32_t value.

vkSequencesCountOffset :: VkDeviceSize

sequencesCountOffset is the byte offset into sequencesCountBuffer where the count value is stored.

vkSequencesIndexBuffer :: VkBuffer

sequencesIndexBuffer must be set if indirectCommandsLayout’s VK_INDIRECT_COMMANDS_LAYOUT_USAGE_INDEXED_SEQUENCES_BIT_NVX is set and provides the used sequence indices as uint32_t array. Otherwise it must be VK_NULL_HANDLE.

vkSequencesIndexOffset :: VkDeviceSize

sequencesIndexOffset is the byte offset into sequencesIndexBuffer where the index values start.

data VkCmdReserveSpaceForCommandsInfoNVX

VkCmdReserveSpaceForCommandsInfoNVX - Structure specifying parameters for the reservation of command buffer space

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_CMD_RESERVE_SPACE_FOR_COMMANDS_INFO_NVX

pNext must be NULL

objectTable must be a valid VkObjectTableNVX handle

indirectCommandsLayout must be a valid VkIndirectCommandsLayoutNVX handle

Both of indirectCommandsLayout, and objectTable must have been created, allocated, or retrieved from the same VkDevice

Host Synchronization

Host access to objectTable must be externally synchronized

See Also

VkIndirectCommandsLayoutNVX, VkObjectTableNVX, VkStructureType, vkCmdReserveSpaceForCommandsNVX

Constructors

VkCmdReserveSpaceForCommandsInfoNVX

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkObjectTable :: VkObjectTableNVX

objectTable is the VkObjectTableNVX to be used for the generation process. Only registered objects at the time vkCmdReserveSpaceForCommandsNVX is called, will be taken into account for the reservation.

vkIndirectCommandsLayout :: VkIndirectCommandsLayoutNVX

indirectCommandsLayout is the VkIndirectCommandsLayoutNVX that must also be used at generation time.

vkMaxSequencesCount :: Word32

maxSequencesCount is the maximum number of sequences for which command buffer space will be reserved.

data VkDeviceGeneratedCommandsFeaturesNVX

VkDeviceGeneratedCommandsFeaturesNVX - Structure specifying physical device support

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType, vkGetPhysicalDeviceGeneratedCommandsPropertiesNVX

Constructors

VkDeviceGeneratedCommandsFeaturesNVX

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DEVICE_GENERATED_COMMANDS_FEATURES_NVX

vkPNext :: Ptr ()

pNext must be NULL

vkComputeBindingPointSupport :: VkBool32

computeBindingPointSupport specifies whether the VkObjectTableNVX supports entries with VK_OBJECT_ENTRY_USAGE_GRAPHICS_BIT_NVX bit set and VkIndirectCommandsLayoutNVX supports VK_PIPELINE_BIND_POINT_COMPUTE.

data VkDeviceGeneratedCommandsLimitsNVX

VkDeviceGeneratedCommandsLimitsNVX - Structure specifying physical device limits

Valid Usage (Implicit)

See Also

VkStructureType, vkGetPhysicalDeviceGeneratedCommandsPropertiesNVX

Constructors

VkDeviceGeneratedCommandsLimitsNVX

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DEVICE_GENERATED_COMMANDS_LIMITS_NVX

vkPNext :: Ptr ()

pNext must be NULL

vkMaxIndirectCommandsLayoutTokenCount :: Word32

maxIndirectCommandsLayoutTokenCount the maximum number of tokens in VkIndirectCommandsLayoutNVX.

vkMaxObjectEntryCounts :: Word32

maxObjectEntryCounts the maximum number of entries per resource type in VkObjectTableNVX.

vkMinSequenceCountBufferOffsetAlignment :: Word32

minSequenceCountBufferOffsetAlignment the minimum alignment for memory addresses optionally used in vkCmdProcessCommandsNVX.

vkMinSequenceIndexBufferOffsetAlignment :: Word32

minSequenceIndexBufferOffsetAlignment the minimum alignment for memory addresses optionally used in vkCmdProcessCommandsNVX.

vkMinCommandsTokenBufferOffsetAlignment :: Word32

minCommandsTokenBufferOffsetAlignment the minimum alignment for memory addresses optionally used in vkCmdProcessCommandsNVX.

data VkIndirectCommandsLayoutCreateInfoNVX

VkIndirectCommandsLayoutCreateInfoNVX - Structure specifying the parameters of a newly created indirect commands layout object

Description

The following code illustrates some of the key flags:

void cmdProcessAllSequences(cmd, objectTable, indirectCommandsLayout, pIndirectCommandsTokens, sequencesCount, indexbuffer, indexbufferoffset)
{
  for (s = 0; s < sequencesCount; s++)
  {
    sequence = s;

    if (indirectCommandsLayout.flags & VK_INDIRECT_COMMANDS_LAYOUT_USAGE_UNORDERED_SEQUENCES_BIT_NVX) {
      sequence = incoherent_implementation_dependent_permutation[ sequence ];
    }
    if (indirectCommandsLayout.flags & VK_INDIRECT_COMMANDS_LAYOUT_USAGE_INDEXED_SEQUENCES_BIT_NVX) {
      sequence = indexbuffer.load_uint32( sequence * sizeof(uint32_t) + indexbufferoffset);
    }

    cmdProcessSequence( cmd, objectTable, indirectCommandsLayout, pIndirectCommandsTokens, sequence );
  }
}

Valid Usage

tokenCount must be greater than 0 and below VkDeviceGeneratedCommandsLimitsNVX::maxIndirectCommandsLayoutTokenCount

If the VkDeviceGeneratedCommandsFeaturesNVX::computeBindingPointSupport feature is not enabled, then pipelineBindPoint must not be VK_PIPELINE_BIND_POINT_COMPUTE

If pTokens contains an entry of VK_INDIRECT_COMMANDS_TOKEN_TYPE_PIPELINE_NVX it must be the first element of the array and there must be only a single element of such token type.

All state binding tokens in pTokens must occur prior work provoking tokens (VK_INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_NVX, VK_INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_INDEXED_NVX, VK_INDIRECT_COMMANDS_TOKEN_TYPE_DISPATCH_NVX).

The content of pTokens must include one single work provoking token that is compatible with the pipelineBindPoint.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_INDIRECT_COMMANDS_LAYOUT_CREATE_INFO_NVX

pNext must be NULL

pipelineBindPoint must be a valid VkPipelineBindPoint value

flags must be a valid combination of VkIndirectCommandsLayoutUsageFlagBitsNVX values

flags must not be 0

pTokens must be a valid pointer to an array of tokenCount valid VkIndirectCommandsLayoutTokenNVX structures

tokenCount must be greater than 0

See Also

VkIndirectCommandsLayoutTokenNVX, VkIndirectCommandsLayoutUsageFlagsNVX, VkPipelineBindPoint, VkStructureType, vkCreateIndirectCommandsLayoutNVX

Constructors

VkIndirectCommandsLayoutCreateInfoNVX

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkPipelineBindPoint :: VkPipelineBindPoint

pipelineBindPoint is the VkPipelineBindPoint that this layout targets.

vkFlags :: VkIndirectCommandsLayoutUsageFlagsNVX

flags is a bitmask of VkIndirectCommandsLayoutUsageFlagBitsNVX specifying usage hints of this layout.

vkTokenCount :: Word32

tokenCount is the length of the individual command sequnce.

vkPTokens :: Ptr VkIndirectCommandsLayoutTokenNVX

pTokens is an array describing each command token in detail. See VkIndirectCommandsTokenTypeNVX and VkIndirectCommandsLayoutTokenNVX below for details.

type VkIndirectCommandsLayoutNVX = Ptr VkIndirectCommandsLayoutNVX_T

VkIndirectCommandsLayoutNVX - Opaque handle to an indirect commands layout object

See Also

VkCmdProcessCommandsInfoNVX, VkCmdReserveSpaceForCommandsInfoNVX, vkCreateIndirectCommandsLayoutNVX, vkDestroyIndirectCommandsLayoutNVX

data VkIndirectCommandsLayoutTokenNVX

VkIndirectCommandsLayoutTokenNVX - Struct specifying the details of an indirect command layout token

Valid Usage (Implicit)

See Also

VkIndirectCommandsLayoutCreateInfoNVX, VkIndirectCommandsTokenTypeNVX

Constructors

VkIndirectCommandsLayoutTokenNVX

Fields

vkTokenType :: VkIndirectCommandsTokenTypeNVX

tokenType must be a valid VkIndirectCommandsTokenTypeNVX value

vkBindingUnit :: Word32

bindingUnit must stay within device supported limits for the appropriate commands.

vkDynamicCount :: Word32

dynamicCount must stay within device supported limits for the appropriate commands.

vkDivisor :: Word32

divisor must be greater than 0 and a power of two.

newtype VkIndirectCommandsLayoutUsageFlagBitsNVX

VkIndirectCommandsLayoutUsageFlagBitsNVX - Bitmask specifying allowed usage of an indirect commands layout

See Also

VkIndirectCommandsLayoutUsageFlagsNVX

Constructors

VkIndirectCommandsLayoutUsageFlagBitsNVX VkFlags

pattern VK_INDIRECT_COMMANDS_LAYOUT_USAGE_UNORDERED_SEQUENCES_BIT_NVX :: VkIndirectCommandsLayoutUsageFlagBitsNVX

VK_INDIRECT_COMMANDS_LAYOUT_USAGE_UNORDERED_SEQUENCES_BIT_NVX specifies that the processing of sequences can happen at an implementation-dependent order, which is not guaranteed to be coherent across multiple invocations.

pattern VK_INDIRECT_COMMANDS_LAYOUT_USAGE_SPARSE_SEQUENCES_BIT_NVX :: VkIndirectCommandsLayoutUsageFlagBitsNVX

VK_INDIRECT_COMMANDS_LAYOUT_USAGE_SPARSE_SEQUENCES_BIT_NVX specifies that there is likely a high difference between allocated number of sequences and actually used.

pattern VK_INDIRECT_COMMANDS_LAYOUT_USAGE_EMPTY_EXECUTIONS_BIT_NVX :: VkIndirectCommandsLayoutUsageFlagBitsNVX

VK_INDIRECT_COMMANDS_LAYOUT_USAGE_EMPTY_EXECUTIONS_BIT_NVX specifies that there are likely many draw or dispatch calls that are zero-sized (zero grid dimension, no primitives to render).

pattern VK_INDIRECT_COMMANDS_LAYOUT_USAGE_INDEXED_SEQUENCES_BIT_NVX :: VkIndirectCommandsLayoutUsageFlagBitsNVX

VK_INDIRECT_COMMANDS_LAYOUT_USAGE_INDEXED_SEQUENCES_BIT_NVX specifies that the input data for the sequences is not implicitly indexed from 0..sequencesUsed but a user provided VkBuffer encoding the index is provided.

type VkIndirectCommandsLayoutUsageFlagsNVX = VkIndirectCommandsLayoutUsageFlagBitsNVX

VkIndirectCommandsLayoutUsageFlagsNVX - Bitmask of VkIndirectCommandsLayoutUsageFlagBitsNVX

Description

VkIndirectCommandsLayoutUsageFlagsNVX is a bitmask type for setting a mask of zero or more VkIndirectCommandsLayoutUsageFlagBitsNVX.

See Also

VkIndirectCommandsLayoutCreateInfoNVX, VkIndirectCommandsLayoutUsageFlagBitsNVX

data VkIndirectCommandsTokenNVX

VkIndirectCommandsTokenNVX - Structure specifying parameters for the reservation of command buffer space

Valid Usage

The buffer’s usage flag must have the VK_BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set.

The offset must be aligned to VkDeviceGeneratedCommandsLimitsNVX::minCommandsTokenBufferOffsetAlignment.

Valid Usage (Implicit)

tokenType must be a valid VkIndirectCommandsTokenTypeNVX value

buffer must be a valid VkBuffer handle

See Also

VkBuffer, VkCmdProcessCommandsInfoNVX, VkDeviceSize, VkIndirectCommandsTokenTypeNVX

Constructors

VkIndirectCommandsTokenNVX

Fields

vkTokenType :: VkIndirectCommandsTokenTypeNVX

tokenType specifies the token command type.

vkBuffer :: VkBuffer

buffer specifies the VkBuffer storing the functional arguments for each squence. These argumetns can be written by the device.

vkOffset :: VkDeviceSize

offset specified an offset into buffer where the arguments start.

newtype VkIndirectCommandsTokenTypeNVX

VkIndirectCommandsTokenTypeNVX - Enum specifying

Description

'

+-----------------------------------------------+----------------------+
| Token type                                    | Equivalent command   |
+===============================================+======================+
| 'VK_INDIRECT_COMMANDS_TOKEN_TYPE_PIPELINE_NVX | 'Graphics.Vulkan.C.C |
| '                                             | ore10.CommandBufferB |
|                                               | uilding.vkCmdBindPip |
|                                               | eline'               |
+-----------------------------------------------+----------------------+
| 'VK_INDIRECT_COMMANDS_TOKEN_TYPE_DESCRIPTOR_S | 'Graphics.Vulkan.C.C |
| ET_NVX'                                       | ore10.CommandBufferB |
|                                               | uilding.vkCmdBindDes |
|                                               | criptorSets'         |
+-----------------------------------------------+----------------------+
| 'VK_INDIRECT_COMMANDS_TOKEN_TYPE_INDEX_BUFFER | 'Graphics.Vulkan.C.C |
| _NVX'                                         | ore10.CommandBufferB |
|                                               | uilding.vkCmdBindInd |
|                                               | exBuffer'            |
+-----------------------------------------------+----------------------+
| 'VK_INDIRECT_COMMANDS_TOKEN_TYPE_VERTEX_BUFFE | 'Graphics.Vulkan.C.C |
| R_NVX'                                        | ore10.CommandBufferB |
|                                               | uilding.vkCmdBindVer |
|                                               | texBuffers'          |
+-----------------------------------------------+----------------------+
| 'VK_INDIRECT_COMMANDS_TOKEN_TYPE_PUSH_CONSTAN | 'Graphics.Vulkan.C.C |
| T_NVX'                                        | ore10.CommandBufferB |
|                                               | uilding.vkCmdPushCon |
|                                               | stants'              |
+-----------------------------------------------+----------------------+
| 'VK_INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_INDEXED | 'Graphics.Vulkan.C.C |
| _NVX'                                         | ore10.CommandBufferB |
|                                               | uilding.vkCmdDrawInd |
|                                               | exedIndirect'        |
+-----------------------------------------------+----------------------+
| 'VK_INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_NVX'    | 'Graphics.Vulkan.C.C |
|                                               | ore10.CommandBufferB |
|                                               | uilding.vkCmdDrawInd |
|                                               | irect'               |
+-----------------------------------------------+----------------------+
| 'VK_INDIRECT_COMMANDS_TOKEN_TYPE_DISPATCH_NVX | 'Graphics.Vulkan.C.C |
| '                                             | ore10.CommandBufferB |
|                                               | uilding.vkCmdDispatc |
|                                               | hIndirect'           |
+-----------------------------------------------+----------------------+

Supported indirect command tokens

See Also

VkIndirectCommandsLayoutTokenNVX, VkIndirectCommandsTokenNVX

Constructors

VkIndirectCommandsTokenTypeNVX Int32

pattern VK_INDIRECT_COMMANDS_TOKEN_TYPE_PIPELINE_NVX :: VkIndirectCommandsTokenTypeNVX

pattern VK_INDIRECT_COMMANDS_TOKEN_TYPE_DESCRIPTOR_SET_NVX :: VkIndirectCommandsTokenTypeNVX

pattern VK_INDIRECT_COMMANDS_TOKEN_TYPE_INDEX_BUFFER_NVX :: VkIndirectCommandsTokenTypeNVX

pattern VK_INDIRECT_COMMANDS_TOKEN_TYPE_VERTEX_BUFFER_NVX :: VkIndirectCommandsTokenTypeNVX

pattern VK_INDIRECT_COMMANDS_TOKEN_TYPE_PUSH_CONSTANT_NVX :: VkIndirectCommandsTokenTypeNVX

pattern VK_INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_INDEXED_NVX :: VkIndirectCommandsTokenTypeNVX

pattern VK_INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_NVX :: VkIndirectCommandsTokenTypeNVX

pattern VK_INDIRECT_COMMANDS_TOKEN_TYPE_DISPATCH_NVX :: VkIndirectCommandsTokenTypeNVX

newtype VkObjectEntryTypeNVX

VkObjectEntryTypeNVX - Enum specifying object table entry type

See Also

VkObjectTableCreateInfoNVX, VkObjectTableDescriptorSetEntryNVX, VkObjectTableEntryNVX, VkObjectTableIndexBufferEntryNVX, VkObjectTablePipelineEntryNVX, VkObjectTablePushConstantEntryNVX, VkObjectTableVertexBufferEntryNVX, vkUnregisterObjectsNVX

Constructors

VkObjectEntryTypeNVX Int32

pattern VK_OBJECT_ENTRY_TYPE_DESCRIPTOR_SET_NVX :: VkObjectEntryTypeNVX

VK_OBJECT_ENTRY_TYPE_DESCRIPTOR_SET_NVX specifies a VkDescriptorSet resource entry that is registered via VkObjectTableDescriptorSetEntryNVX.

pattern VK_OBJECT_ENTRY_TYPE_PIPELINE_NVX :: VkObjectEntryTypeNVX

VK_OBJECT_ENTRY_TYPE_PIPELINE_NVX specifies a VkPipeline resource entry that is registered via VkObjectTablePipelineEntryNVX.

pattern VK_OBJECT_ENTRY_TYPE_INDEX_BUFFER_NVX :: VkObjectEntryTypeNVX

VK_OBJECT_ENTRY_TYPE_INDEX_BUFFER_NVX specifies a VkBuffer resource entry that is registered via VkObjectTableIndexBufferEntryNVX.

pattern VK_OBJECT_ENTRY_TYPE_VERTEX_BUFFER_NVX :: VkObjectEntryTypeNVX

VK_OBJECT_ENTRY_TYPE_VERTEX_BUFFER_NVX specifies a VkBuffer resource entry that is registered via VkObjectTableVertexBufferEntryNVX.

pattern VK_OBJECT_ENTRY_TYPE_PUSH_CONSTANT_NVX :: VkObjectEntryTypeNVX

VK_OBJECT_ENTRY_TYPE_PUSH_CONSTANT_NVX specifies the resource entry is registered via VkObjectTablePushConstantEntryNVX.

newtype VkObjectEntryUsageFlagBitsNVX

VkObjectEntryUsageFlagBitsNVX - Bitmask specifying allowed usage of an object entry

See Also

VkObjectEntryUsageFlagsNVX

Constructors

VkObjectEntryUsageFlagBitsNVX VkFlags

pattern VK_OBJECT_ENTRY_USAGE_GRAPHICS_BIT_NVX :: VkObjectEntryUsageFlagBitsNVX

VK_OBJECT_ENTRY_USAGE_GRAPHICS_BIT_NVX specifies that the resource is bound to VK_PIPELINE_BIND_POINT_GRAPHICS

pattern VK_OBJECT_ENTRY_USAGE_COMPUTE_BIT_NVX :: VkObjectEntryUsageFlagBitsNVX

VK_OBJECT_ENTRY_USAGE_COMPUTE_BIT_NVX specifies that the resource is bound to VK_PIPELINE_BIND_POINT_COMPUTE

type VkObjectEntryUsageFlagsNVX = VkObjectEntryUsageFlagBitsNVX

VkObjectEntryUsageFlagsNVX - Bitmask of VkObjectEntryUsageFlagBitsNVX

Description

VkObjectEntryUsageFlagsNVX is a bitmask type for setting a mask of zero or more VkObjectEntryUsageFlagBitsNVX.

See Also

VkObjectEntryUsageFlagBitsNVX, VkObjectTableCreateInfoNVX, VkObjectTableDescriptorSetEntryNVX, VkObjectTableEntryNVX, VkObjectTableIndexBufferEntryNVX, VkObjectTablePipelineEntryNVX, VkObjectTablePushConstantEntryNVX, VkObjectTableVertexBufferEntryNVX

data VkObjectTableCreateInfoNVX

VkObjectTableCreateInfoNVX - Structure specifying the parameters of a newly created object table

Valid Usage

If the VkDeviceGeneratedCommandsFeaturesNVX::computeBindingPointSupport feature is not enabled, pObjectEntryUsageFlags must not contain VK_OBJECT_ENTRY_USAGE_COMPUTE_BIT_NVX

Any value within pObjectEntryCounts must not exceed VkDeviceGeneratedCommandsLimitsNVX::maxObjectEntryCounts

maxUniformBuffersPerDescriptor must be within the limits supported by the device.

maxStorageBuffersPerDescriptor must be within the limits supported by the device.

maxStorageImagesPerDescriptor must be within the limits supported by the device.

maxSampledImagesPerDescriptor must be within the limits supported by the device.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_OBJECT_TABLE_CREATE_INFO_NVX

pNext must be NULL

pObjectEntryTypes must be a valid pointer to an array of objectCount valid VkObjectEntryTypeNVX values

pObjectEntryCounts must be a valid pointer to an array of objectCount uint32_t values

pObjectEntryUsageFlags must be a valid pointer to an array of objectCount valid combinations of VkObjectEntryUsageFlagBitsNVX values

Each element of pObjectEntryUsageFlags must not be 0

objectCount must be greater than 0

See Also

VkObjectEntryTypeNVX, VkObjectEntryUsageFlagsNVX, VkStructureType, vkCreateObjectTableNVX

Constructors

VkObjectTableCreateInfoNVX

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkObjectCount :: Word32

objectCount is the number of entry configurations that the object table supports.

vkPObjectEntryTypes :: Ptr VkObjectEntryTypeNVX

pObjectEntryTypes is an array of VkObjectEntryTypeNVX values providing the entry type of a given configuration.

vkPObjectEntryCounts :: Ptr Word32

pObjectEntryCounts is an array of counts of how many objects can be registered in the table.

vkPObjectEntryUsageFlags :: Ptr VkObjectEntryUsageFlagsNVX

pObjectEntryUsageFlags is an array of bitmasks of VkObjectEntryUsageFlagBitsNVX specifying the binding usage of the entry.

vkMaxUniformBuffersPerDescriptor :: Word32

maxUniformBuffersPerDescriptor is the maximum number of VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER or VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC used by any single registered VkDescriptorSet in this table.

vkMaxStorageBuffersPerDescriptor :: Word32

maxStorageBuffersPerDescriptor is the maximum number of VK_DESCRIPTOR_TYPE_STORAGE_BUFFER or VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC used by any single registered VkDescriptorSet in this table.

vkMaxStorageImagesPerDescriptor :: Word32

maxStorageImagesPerDescriptor is the maximum number of VK_DESCRIPTOR_TYPE_STORAGE_IMAGE or VK_DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER used by any single registered VkDescriptorSet in this table.

vkMaxSampledImagesPerDescriptor :: Word32

maxSampledImagesPerDescriptor is the maximum number of VK_DESCRIPTOR_TYPE_SAMPLER, VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, VK_DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER or VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT used by any single registered VkDescriptorSet in this table.

vkMaxPipelineLayouts :: Word32

maxPipelineLayouts is the maximum number of unique VkPipelineLayout used by any registered VkDescriptorSet or VkPipeline in this table.

data VkObjectTableDescriptorSetEntryNVX

VkObjectTableDescriptorSetEntryNVX - Parameters of an object table descriptor set entry

Valid Usage

type must be VK_OBJECT_ENTRY_TYPE_DESCRIPTOR_SET_NVX

Valid Usage (Implicit)

type must be a valid VkObjectEntryTypeNVX value

flags must be a valid combination of VkObjectEntryUsageFlagBitsNVX values

flags must not be 0

pipelineLayout must be a valid VkPipelineLayout handle

descriptorSet must be a valid VkDescriptorSet handle

Both of descriptorSet, and pipelineLayout must have been created, allocated, or retrieved from the same VkDevice

See Also

VkDescriptorSet, VkObjectEntryTypeNVX, VkObjectEntryUsageFlagsNVX, VkPipelineLayout

Constructors

VkObjectTableDescriptorSetEntryNVX

Fields

vkType :: VkObjectEntryTypeNVX

vkFlags :: VkObjectEntryUsageFlagsNVX

vkPipelineLayout :: VkPipelineLayout

pipelineLayout specifies the VkPipelineLayout that the descriptorSet is used with.

vkDescriptorSet :: VkDescriptorSet

descriptorSet specifies the VkDescriptorSet that can be bound with this entry.

data VkObjectTableEntryNVX

VkObjectTableEntryNVX - Common parameters of an object table resource entry

Valid Usage

If the VkDeviceGeneratedCommandsFeaturesNVX::computeBindingPointSupport feature is not enabled, flags must not contain VK_OBJECT_ENTRY_USAGE_COMPUTE_BIT_NVX

Valid Usage (Implicit)

type must be a valid VkObjectEntryTypeNVX value

flags must be a valid combination of VkObjectEntryUsageFlagBitsNVX values

flags must not be 0

See Also

VkObjectEntryTypeNVX, VkObjectEntryUsageFlagsNVX, vkRegisterObjectsNVX

Constructors

VkObjectTableEntryNVX

Fields

vkType :: VkObjectEntryTypeNVX

type defines the entry type

vkFlags :: VkObjectEntryUsageFlagsNVX

flags defines which VkPipelineBindPoint the resource can be used with. Some entry types allow only a single flag to be set.

data VkObjectTableIndexBufferEntryNVX

VkObjectTableIndexBufferEntryNVX - Parameters of an object table index buffer entry

Valid Usage (Implicit)

See Also

VkBuffer, VkIndexType, VkObjectEntryTypeNVX, VkObjectEntryUsageFlagsNVX

Constructors

VkObjectTableIndexBufferEntryNVX

Fields

vkType :: VkObjectEntryTypeNVX

type must be a valid VkObjectEntryTypeNVX value

vkFlags :: VkObjectEntryUsageFlagsNVX

flags must not be 0

vkBuffer :: VkBuffer

buffer must be a valid VkBuffer handle

vkIndexType :: VkIndexType

indexType must be a valid VkIndexType value

type VkObjectTableNVX = Ptr VkObjectTableNVX_T

VkObjectTableNVX - Opaque handle to an object table

See Also

VkCmdProcessCommandsInfoNVX, VkCmdReserveSpaceForCommandsInfoNVX, vkCreateObjectTableNVX, vkDestroyObjectTableNVX, vkRegisterObjectsNVX, vkUnregisterObjectsNVX

data VkObjectTablePipelineEntryNVX

VkObjectTablePipelineEntryNVX - Parameters of an object table pipeline entry

Valid Usage (Implicit)

See Also

VkObjectEntryTypeNVX, VkObjectEntryUsageFlagsNVX, VkPipeline

Constructors

VkObjectTablePipelineEntryNVX

Fields

vkType :: VkObjectEntryTypeNVX

type must be a valid VkObjectEntryTypeNVX value

vkFlags :: VkObjectEntryUsageFlagsNVX

flags must not be 0

vkPipeline :: VkPipeline

pipeline must be a valid VkPipeline handle

data VkObjectTablePushConstantEntryNVX

VkObjectTablePushConstantEntryNVX - Parameters of an object table push constant entry

Valid Usage (Implicit)

See Also

VkObjectEntryTypeNVX, VkObjectEntryUsageFlagsNVX, VkPipelineLayout, VkShaderStageFlags

Constructors

VkObjectTablePushConstantEntryNVX

Fields

vkType :: VkObjectEntryTypeNVX

type must be a valid VkObjectEntryTypeNVX value

vkFlags :: VkObjectEntryUsageFlagsNVX

flags must not be 0

vkPipelineLayout :: VkPipelineLayout

pipelineLayout must be a valid VkPipelineLayout handle

vkStageFlags :: VkShaderStageFlags

stageFlags must not be 0

data VkObjectTableVertexBufferEntryNVX

VkObjectTableVertexBufferEntryNVX - Parameters of an object table vertex buffer entry

Valid Usage (Implicit)

See Also

VkBuffer, VkObjectEntryTypeNVX, VkObjectEntryUsageFlagsNVX

Constructors

VkObjectTableVertexBufferEntryNVX

Fields

vkType :: VkObjectEntryTypeNVX

type must be a valid VkObjectEntryTypeNVX value

vkFlags :: VkObjectEntryUsageFlagsNVX

flags must not be 0

vkBuffer :: VkBuffer

buffer must be a valid VkBuffer handle

type FN_vkCmdProcessCommandsNVX = ("commandBuffer" ::: VkCommandBuffer) -> ("pProcessCommandsInfo" ::: Ptr VkCmdProcessCommandsInfoNVX) -> IO ()

type PFN_vkCmdProcessCommandsNVX = FunPtr FN_vkCmdProcessCommandsNVX

vkCmdProcessCommandsNVX :: DeviceCmds -> ("commandBuffer" ::: VkCommandBuffer) -> ("pProcessCommandsInfo" ::: Ptr VkCmdProcessCommandsInfoNVX) -> IO ()

vkCmdProcessCommandsNVX - Performs the generation of commands on the device

Parameters

commandBuffer is the primary command buffer in which the generation process takes space.

pProcessCommandsInfo is a pointer to an instance of the VkCmdProcessCommandsInfoNVX structure containing parameters affecting the processing of commands.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pProcessCommandsInfo must be a valid pointer to a valid VkCmdProcessCommandsInfoNVX structure

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

This command must only be called inside of a render pass instance

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Inside          | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCmdProcessCommandsInfoNVX, VkCommandBuffer

type FN_vkCmdReserveSpaceForCommandsNVX = ("commandBuffer" ::: VkCommandBuffer) -> ("pReserveSpaceInfo" ::: Ptr VkCmdReserveSpaceForCommandsInfoNVX) -> IO ()

type PFN_vkCmdReserveSpaceForCommandsNVX = FunPtr FN_vkCmdReserveSpaceForCommandsNVX

vkCmdReserveSpaceForCommandsNVX :: DeviceCmds -> ("commandBuffer" ::: VkCommandBuffer) -> ("pReserveSpaceInfo" ::: Ptr VkCmdReserveSpaceForCommandsInfoNVX) -> IO ()

vkCmdReserveSpaceForCommandsNVX - Perform a reservation of command buffer space

Parameters

commandBuffer is the secondary command buffer in which the space for device-generated commands is reserved.

pProcessCommandsInfo is a pointer to an instance of the VkCmdReserveSpaceForCommandsInfoNVX structure containing parameters affecting the reservation of command buffer space.

Valid Usage

The provided commandBuffer must not have had a prior space reservation since its creation or the last reset.

The state of the commandBuffer must be legal to execute all commands within the sequence provided by the indirectCommandsLayout member of pProcessCommandsInfo.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pReserveSpaceInfo must be a valid pointer to a valid VkCmdReserveSpaceForCommandsInfoNVX structure

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

This command must only be called inside of a render pass instance

commandBuffer must be a secondary VkCommandBuffer

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Secondary       | Inside          | Graphics        |                 |
|                 |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCmdReserveSpaceForCommandsInfoNVX, VkCommandBuffer

type FN_vkCreateIndirectCommandsLayoutNVX = ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkIndirectCommandsLayoutCreateInfoNVX) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pIndirectCommandsLayout" ::: Ptr VkIndirectCommandsLayoutNVX) -> IO VkResult

type PFN_vkCreateIndirectCommandsLayoutNVX = FunPtr FN_vkCreateIndirectCommandsLayoutNVX

vkCreateIndirectCommandsLayoutNVX :: DeviceCmds -> ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkIndirectCommandsLayoutCreateInfoNVX) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pIndirectCommandsLayout" ::: Ptr VkIndirectCommandsLayoutNVX) -> IO VkResult

vkCreateIndirectCommandsLayoutNVX - Create an indirect command layout object

Parameters

device is the logical device that creates the indirect command layout.

pCreateInfo is a pointer to an instance of the VkIndirectCommandsLayoutCreateInfoNVX structure containing parameters affecting creation of the indirect command layout.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pIndirectCommandsLayout points to a VkIndirectCommandsLayoutNVX handle in which the resulting indirect command layout is returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkIndirectCommandsLayoutCreateInfoNVX structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pIndirectCommandsLayout must be a valid pointer to a VkIndirectCommandsLayoutNVX handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDevice, VkIndirectCommandsLayoutCreateInfoNVX, VkIndirectCommandsLayoutNVX

type FN_vkCreateObjectTableNVX = ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkObjectTableCreateInfoNVX) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pObjectTable" ::: Ptr VkObjectTableNVX) -> IO VkResult

type PFN_vkCreateObjectTableNVX = FunPtr FN_vkCreateObjectTableNVX

vkCreateObjectTableNVX :: DeviceCmds -> ("device" ::: VkDevice) -> ("pCreateInfo" ::: Ptr VkObjectTableCreateInfoNVX) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> ("pObjectTable" ::: Ptr VkObjectTableNVX) -> IO VkResult

vkCreateObjectTableNVX - Create an object table

Parameters

device is the logical device that creates the object table.

pCreateInfo is a pointer to an instance of the VkObjectTableCreateInfoNVX structure containing parameters affecting creation of the table.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pObjectTable points to a VkObjectTableNVX handle in which the resulting object table is returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkObjectTableCreateInfoNVX structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pObjectTable must be a valid pointer to a VkObjectTableNVX handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDevice, VkObjectTableCreateInfoNVX, VkObjectTableNVX

type FN_vkDestroyIndirectCommandsLayoutNVX = ("device" ::: VkDevice) -> ("indirectCommandsLayout" ::: VkIndirectCommandsLayoutNVX) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> IO ()

type PFN_vkDestroyIndirectCommandsLayoutNVX = FunPtr FN_vkDestroyIndirectCommandsLayoutNVX

vkDestroyIndirectCommandsLayoutNVX :: DeviceCmds -> ("device" ::: VkDevice) -> ("indirectCommandsLayout" ::: VkIndirectCommandsLayoutNVX) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> IO ()

vkDestroyIndirectCommandsLayoutNVX - Destroy an object table

Parameters

device is the logical device that destroys the layout.

indirectCommandsLayout is the table to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to indirectCommandsLayout must have completed execution

If VkAllocationCallbacks were provided when objectTable was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when objectTable was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid VkDevice handle

indirectCommandsLayout must be a valid VkIndirectCommandsLayoutNVX handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

indirectCommandsLayout must have been created, allocated, or retrieved from device

See Also

VkAllocationCallbacks, VkDevice, VkIndirectCommandsLayoutNVX

type FN_vkDestroyObjectTableNVX = ("device" ::: VkDevice) -> ("objectTable" ::: VkObjectTableNVX) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> IO ()

type PFN_vkDestroyObjectTableNVX = FunPtr FN_vkDestroyObjectTableNVX

vkDestroyObjectTableNVX :: DeviceCmds -> ("device" ::: VkDevice) -> ("objectTable" ::: VkObjectTableNVX) -> ("pAllocator" ::: Ptr VkAllocationCallbacks) -> IO ()

vkDestroyObjectTableNVX - Destroy an object table

Parameters

device is the logical device that destroys the table.

objectTable is the table to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to objectTable must have completed execution.

If VkAllocationCallbacks were provided when objectTable was created, a compatible set of callbacks must be provided here.

If no VkAllocationCallbacks were provided when objectTable was created, pAllocator must be NULL.

Valid Usage (Implicit)

device must be a valid VkDevice handle

objectTable must be a valid VkObjectTableNVX handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

objectTable must have been created, allocated, or retrieved from device

Host Synchronization

Host access to objectTable must be externally synchronized

See Also

VkAllocationCallbacks, VkDevice, VkObjectTableNVX

type FN_vkGetPhysicalDeviceGeneratedCommandsPropertiesNVX = ("physicalDevice" ::: VkPhysicalDevice) -> ("pFeatures" ::: Ptr VkDeviceGeneratedCommandsFeaturesNVX) -> ("pLimits" ::: Ptr VkDeviceGeneratedCommandsLimitsNVX) -> IO ()

type PFN_vkGetPhysicalDeviceGeneratedCommandsPropertiesNVX = FunPtr FN_vkGetPhysicalDeviceGeneratedCommandsPropertiesNVX

vkGetPhysicalDeviceGeneratedCommandsPropertiesNVX :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pFeatures" ::: Ptr VkDeviceGeneratedCommandsFeaturesNVX) -> ("pLimits" ::: Ptr VkDeviceGeneratedCommandsLimitsNVX) -> IO ()

vkGetPhysicalDeviceGeneratedCommandsPropertiesNVX - Returns device-generated commands related properties of a physical device

Parameters

physicalDevice is the handle to the physical device whose properties will be queried.

pFeatures points to an instance of the VkDeviceGeneratedCommandsFeaturesNVX structure, that will be filled with returned information.

pLimits points to an instance of the VkDeviceGeneratedCommandsLimitsNVX structure, that will be filled with returned information.

Valid Usage (Implicit)

See Also

VkDeviceGeneratedCommandsFeaturesNVX, VkDeviceGeneratedCommandsLimitsNVX, VkPhysicalDevice

type FN_vkRegisterObjectsNVX = ("device" ::: VkDevice) -> ("objectTable" ::: VkObjectTableNVX) -> ("objectCount" ::: Word32) -> ("ppObjectTableEntries" ::: Ptr (Ptr VkObjectTableEntryNVX)) -> ("pObjectIndices" ::: Ptr Word32) -> IO VkResult

type PFN_vkRegisterObjectsNVX = FunPtr FN_vkRegisterObjectsNVX

vkRegisterObjectsNVX :: DeviceCmds -> ("device" ::: VkDevice) -> ("objectTable" ::: VkObjectTableNVX) -> ("objectCount" ::: Word32) -> ("ppObjectTableEntries" ::: Ptr (Ptr VkObjectTableEntryNVX)) -> ("pObjectIndices" ::: Ptr Word32) -> IO VkResult

vkRegisterObjectsNVX - Register resource bindings in an object table

Parameters

device is the logical device that creates the object table.

objectTable is the table for which the resources are registered.

objectCount is the number of resources to register.

ppObjectTableEntries provides an array for detailed binding informations, each array element is a pointer to a struct of type VkObjectTablePipelineEntryNVX, VkObjectTableDescriptorSetEntryNVX, VkObjectTableVertexBufferEntryNVX, VkObjectTableIndexBufferEntryNVX or VkObjectTablePushConstantEntryNVX (see below for details).

pObjectIndices are the indices at which each resource is registered.

Valid Usage

The contents of pObjectTableEntry must yield plausible bindings supported by the device.

At any pObjectIndices there must not be a registered resource already.

Any value inside pObjectIndices must be below the appropriate VkObjectTableCreateInfoNVX::pObjectEntryCounts limits provided at objectTable creation time.

Valid Usage (Implicit)

device must be a valid VkDevice handle

objectTable must be a valid VkObjectTableNVX handle

ppObjectTableEntries must be a valid pointer to an array of objectCount valid VkObjectTableEntryNVX structures

pObjectIndices must be a valid pointer to an array of objectCount uint32_t values

objectCount must be greater than 0

objectTable must have been created, allocated, or retrieved from device

Host Synchronization

Host access to objectTable must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDevice, VkObjectTableEntryNVX, VkObjectTableNVX

type FN_vkUnregisterObjectsNVX = ("device" ::: VkDevice) -> ("objectTable" ::: VkObjectTableNVX) -> ("objectCount" ::: Word32) -> ("pObjectEntryTypes" ::: Ptr VkObjectEntryTypeNVX) -> ("pObjectIndices" ::: Ptr Word32) -> IO VkResult

type PFN_vkUnregisterObjectsNVX = FunPtr FN_vkUnregisterObjectsNVX

vkUnregisterObjectsNVX :: DeviceCmds -> ("device" ::: VkDevice) -> ("objectTable" ::: VkObjectTableNVX) -> ("objectCount" ::: Word32) -> ("pObjectEntryTypes" ::: Ptr VkObjectEntryTypeNVX) -> ("pObjectIndices" ::: Ptr Word32) -> IO VkResult

vkUnregisterObjectsNVX - Unregister resource bindings in an object table

Parameters

device is the logical device that creates the object table.

objectTable is the table from which the resources are unregistered.

objectCount is the number of resources being removed from the object table.

pObjectEntryType provides an array of VkObjectEntryTypeNVX for the resources being removed.

pObjectIndices provides the array of object indices to be removed.

Valid Usage

At any pObjectIndices there must be a registered resource already.

The pObjectEntryTypes of the resource at pObjectIndices must match.

All operations on the device using the registered resource must have been completed.

Valid Usage (Implicit)

device must be a valid VkDevice handle

objectTable must be a valid VkObjectTableNVX handle

pObjectEntryTypes must be a valid pointer to an array of objectCount valid VkObjectEntryTypeNVX values

pObjectIndices must be a valid pointer to an array of objectCount uint32_t values

objectCount must be greater than 0

objectTable must have been created, allocated, or retrieved from device

Host Synchronization

Host access to objectTable must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDevice, VkObjectEntryTypeNVX, VkObjectTableNVX

pattern VK_ACCESS_COMMAND_PROCESS_READ_BIT_NVX :: VkAccessFlagBits

pattern VK_ACCESS_COMMAND_PROCESS_WRITE_BIT_NVX :: VkAccessFlagBits

pattern VK_NVX_DEVICE_GENERATED_COMMANDS_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_NVX_DEVICE_GENERATED_COMMANDS_SPEC_VERSION :: Integral a => a

pattern VK_OBJECT_TYPE_INDIRECT_COMMANDS_LAYOUT_NVX :: VkObjectType

pattern VK_OBJECT_TYPE_OBJECT_TABLE_NVX :: VkObjectType

pattern VK_PIPELINE_STAGE_COMMAND_PROCESS_BIT_NVX :: VkPipelineStageFlagBits

pattern VK_STRUCTURE_TYPE_CMD_PROCESS_COMMANDS_INFO_NVX :: VkStructureType

pattern VK_STRUCTURE_TYPE_CMD_RESERVE_SPACE_FOR_COMMANDS_INFO_NVX :: VkStructureType

pattern VK_STRUCTURE_TYPE_DEVICE_GENERATED_COMMANDS_FEATURES_NVX :: VkStructureType

pattern VK_STRUCTURE_TYPE_DEVICE_GENERATED_COMMANDS_LIMITS_NVX :: VkStructureType

pattern VK_STRUCTURE_TYPE_INDIRECT_COMMANDS_LAYOUT_CREATE_INFO_NVX :: VkStructureType

pattern VK_STRUCTURE_TYPE_OBJECT_TABLE_CREATE_INFO_NVX :: VkStructureType