Graphics.Vulkan.C.Extensions.VK_NV_clip_space_w_scaling

Documentation

data VkPipelineViewportWScalingStateCreateInfoNV

VkPipelineViewportWScalingStateCreateInfoNV - Structure specifying parameters of a newly created pipeline viewport W scaling state

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType, VkViewportWScalingNV

Constructors

VkPipelineViewportWScalingStateCreateInfoNV

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_W_SCALING_STATE_CREATE_INFO_NV

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkViewportWScalingEnable :: VkBool32

viewportWScalingEnable controls whether viewport W scaling is enabled.

vkViewportCount :: Word32

viewportCount must be greater than 0

vkPViewportWScalings :: Ptr VkViewportWScalingNV

pViewportWScalings is a pointer to an array of VkViewportWScalingNV structures, which define the W scaling parameters for the corresponding viewport. If the viewport W scaling state is dynamic, this member is ignored.

data VkViewportWScalingNV

VkViewportWScalingNV - Structure specifying a viewport

See Also

VkPipelineViewportWScalingStateCreateInfoNV, vkCmdSetViewportWScalingNV

Constructors

VkViewportWScalingNV

Fields

vkXcoeff :: CFloat

xcoeff and ycoeff are the viewport’s W scaling factor for x and y respectively.

vkYcoeff :: CFloat

type FN_vkCmdSetViewportWScalingNV = ("commandBuffer" ::: VkCommandBuffer) -> ("firstViewport" ::: Word32) -> ("viewportCount" ::: Word32) -> ("pViewportWScalings" ::: Ptr VkViewportWScalingNV) -> IO ()

type PFN_vkCmdSetViewportWScalingNV = FunPtr FN_vkCmdSetViewportWScalingNV

vkCmdSetViewportWScalingNV :: DeviceCmds -> ("commandBuffer" ::: VkCommandBuffer) -> ("firstViewport" ::: Word32) -> ("viewportCount" ::: Word32) -> ("pViewportWScalings" ::: Ptr VkViewportWScalingNV) -> IO ()

vkCmdSetViewportWScalingNV - Set the viewport W scaling on a command buffer

Parameters

commandBuffer is the command buffer into which the command will be recorded.

firstViewport is the index of the first viewport whose parameters are updated by the command.

viewportCount is the number of viewports whose parameters are updated by the command.

pViewportWScalings is a pointer to an array of VkViewportWScalingNV structures specifying viewport parameters.

Description

The viewport parameters taken from element i of pViewportWScalings replace the current state for the viewport index firstViewport + i, for i in [0, viewportCount).

Valid Usage

The bound graphics pipeline must have been created with the VK_DYNAMIC_STATE_VIEWPORT_W_SCALING_NV dynamic state enabled

firstViewport must be less than VkPhysicalDeviceLimits::maxViewports

The sum of firstViewport and viewportCount must be between 1 and VkPhysicalDeviceLimits::maxViewports, inclusive

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pViewportWScalings must be a valid pointer to an array of viewportCount VkViewportWScalingNV structures

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics operations

viewportCount must be greater than 0

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkViewportWScalingNV

pattern VK_DYNAMIC_STATE_VIEWPORT_W_SCALING_NV :: VkDynamicState

pattern VK_NV_CLIP_SPACE_W_SCALING_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_NV_CLIP_SPACE_W_SCALING_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_W_SCALING_STATE_CREATE_INFO_NV :: VkStructureType