Graphics.Vulkan.C.Extensions.VK_NV_compute_shader_derivatives

Documentation

data VkPhysicalDeviceComputeShaderDerivativesFeaturesNV

VkPhysicalDeviceComputeShaderDerivativesFeaturesNV - Structure describing compute shader derivative features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceComputeShaderDerivativesFeaturesNV structure describe the following features:

Description

See Compute Shader Derivatives for more information.

If the VkPhysicalDeviceComputeShaderDerivativesFeaturesNV structure is included in the pNext chain of VkPhysicalDeviceFeatures2KHR, it is filled with values indicating whether each feature is supported. VkPhysicalDeviceComputeShaderDerivativesFeaturesNV can also be used in the pNext chain of VkDeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

VkPhysicalDeviceComputeShaderDerivativesFeaturesNV

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_COMPUTE_SHADER_DERIVATIVES_FEATURES_NV

vkPNext :: Ptr ()

vkComputeDerivativeGroupQuads :: VkBool32

computeDerivativeGroupQuads indicates that the implementation supports the ComputeDerivativeGroupQuadsNV SPIR-V capability.

vkComputeDerivativeGroupLinear :: VkBool32

computeDerivativeGroupLinear indicates that the implementation supports the ComputeDerivativeGroupLinearNV SPIR-V capability.

pattern VK_NV_COMPUTE_SHADER_DERIVATIVES_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_NV_COMPUTE_SHADER_DERIVATIVES_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_COMPUTE_SHADER_DERIVATIVES_FEATURES_NV :: VkStructureType