Graphics.Vulkan.C.Extensions.VK_NV_cooperative_matrix

Documentation

newtype VkComponentTypeNV

VkComponentTypeNV - Specify SPIR-V cooperative matrix component type

See Also

VkCooperativeMatrixPropertiesNV

Constructors

VkComponentTypeNV Int32

pattern VK_COMPONENT_TYPE_FLOAT16_NV :: VkComponentTypeNV

VK_COMPONENT_TYPE_FLOAT16_NV corresponds to SPIR-V OpTypeFloat 16.

pattern VK_COMPONENT_TYPE_FLOAT32_NV :: VkComponentTypeNV

VK_COMPONENT_TYPE_FLOAT32_NV corresponds to SPIR-V OpTypeFloat 32.

pattern VK_COMPONENT_TYPE_FLOAT64_NV :: VkComponentTypeNV

VK_COMPONENT_TYPE_FLOAT64_NV corresponds to SPIR-V OpTypeFloat 64.

pattern VK_COMPONENT_TYPE_SINT8_NV :: VkComponentTypeNV

VK_COMPONENT_TYPE_SINT8_NV corresponds to SPIR-V OpTypeInt 8 1.

pattern VK_COMPONENT_TYPE_SINT16_NV :: VkComponentTypeNV

VK_COMPONENT_TYPE_SINT16_NV corresponds to SPIR-V OpTypeInt 16 1.

pattern VK_COMPONENT_TYPE_SINT32_NV :: VkComponentTypeNV

VK_COMPONENT_TYPE_SINT32_NV corresponds to SPIR-V OpTypeInt 32 1.

pattern VK_COMPONENT_TYPE_SINT64_NV :: VkComponentTypeNV

VK_COMPONENT_TYPE_SINT64_NV corresponds to SPIR-V OpTypeInt 64 1.

pattern VK_COMPONENT_TYPE_UINT8_NV :: VkComponentTypeNV

VK_COMPONENT_TYPE_UINT8_NV corresponds to SPIR-V OpTypeInt 8 0.

pattern VK_COMPONENT_TYPE_UINT16_NV :: VkComponentTypeNV

VK_COMPONENT_TYPE_UINT16_NV corresponds to SPIR-V OpTypeInt 16 0.

pattern VK_COMPONENT_TYPE_UINT32_NV :: VkComponentTypeNV

VK_COMPONENT_TYPE_UINT32_NV corresponds to SPIR-V OpTypeInt 32 0.

pattern VK_COMPONENT_TYPE_UINT64_NV :: VkComponentTypeNV

VK_COMPONENT_TYPE_UINT64_NV corresponds to SPIR-V OpTypeInt 64 0.

data VkCooperativeMatrixPropertiesNV

VkCooperativeMatrixPropertiesNV - Structure specifying cooperative matrix properties

Description

If some types are preferred over other types (e.g. for performance), they should appear earlier in the list enumerated by vkGetPhysicalDeviceCooperativeMatrixPropertiesNV.

At least one entry in the list must have power of two values for all of MSize, KSize, and NSize.

Valid Usage (Implicit)

See Also

VkComponentTypeNV, VkScopeNV, VkStructureType, vkGetPhysicalDeviceCooperativeMatrixPropertiesNV

Constructors

VkCooperativeMatrixPropertiesNV

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_COOPERATIVE_MATRIX_PROPERTIES_NV

vkPNext :: Ptr ()

pNext must be NULL

vkMSize :: Word32

MSize is the number of rows in matrices A, C, and D.

vkNSize :: Word32

NSize is the number of columns in matrices B, C, D.

vkKSize :: Word32

KSize is the number of columns in matrix A and rows in matrix B.

vkAType :: VkComponentTypeNV

AType must be a valid VkComponentTypeNV value

vkBType :: VkComponentTypeNV

BType must be a valid VkComponentTypeNV value

vkCType :: VkComponentTypeNV

CType must be a valid VkComponentTypeNV value

vkDType :: VkComponentTypeNV

DType must be a valid VkComponentTypeNV value

vkScope :: VkScopeNV

scope must be a valid VkScopeNV value

data VkPhysicalDeviceCooperativeMatrixFeaturesNV

VkPhysicalDeviceCooperativeMatrixFeaturesNV - Structure describing cooperative matrix features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceCooperativeMatrixFeaturesNV structure describe the following features:

Description

If the VkPhysicalDeviceCooperativeMatrixFeaturesNV structure is included in the pNext chain of VkPhysicalDeviceFeatures2KHR, it is filled with values indicating whether the feature is supported. VkPhysicalDeviceCooperativeMatrixFeaturesNV can also be used in the pNext chain of VkDeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

VkPhysicalDeviceCooperativeMatrixFeaturesNV

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_COOPERATIVE_MATRIX_FEATURES_NV

vkPNext :: Ptr ()

vkCooperativeMatrix :: VkBool32

cooperativeMatrix indicates that the implementation supports the CooperativeMatrixNV SPIR-V capability.

vkCooperativeMatrixRobustBufferAccess :: VkBool32

cooperativeMatrixRobustBufferAccess indicates that the implementation supports robust buffer access for SPIR-V OpCooperativeMatrixLoadNV and OpCooperativeMatrixStoreNV instructions.

data VkPhysicalDeviceCooperativeMatrixPropertiesNV

VkPhysicalDeviceCooperativeMatrixPropertiesNV - Structure describing cooperative matrix properties supported by an implementation

Members

The members of the VkPhysicalDeviceCooperativeMatrixPropertiesNV structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceCooperativeMatrixPropertiesNV structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkShaderStageFlags, VkStructureType

Constructors

VkPhysicalDeviceCooperativeMatrixPropertiesNV

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_COOPERATIVE_MATRIX_PROPERTIES_NV

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkCooperativeMatrixSupportedStages :: VkShaderStageFlags

cooperativeMatrixSupportedStages is a bitfield of VkShaderStageFlagBits describing the shader stages that cooperative matrix instructions are supported in. cooperativeMatrixSupportedStages will have the VK_SHADER_STAGE_COMPUTE_BIT bit set if any of the physical device’s queues support VK_QUEUE_COMPUTE_BIT.

newtype VkScopeNV

VkScopeNV - Specify SPIR-V scope

Description

All enum values match the corresponding SPIR-V value.

See Also

VkCooperativeMatrixPropertiesNV

Constructors

VkScopeNV Int32

pattern VK_SCOPE_DEVICE_NV :: VkScopeNV

VK_SCOPE_DEVICE_NV corresponds to SPIR-V Device scope.

pattern VK_SCOPE_WORKGROUP_NV :: VkScopeNV

VK_SCOPE_WORKGROUP_NV corresponds to SPIR-V Workgroup scope.

pattern VK_SCOPE_SUBGROUP_NV :: VkScopeNV

VK_SCOPE_SUBGROUP_NV corresponds to SPIR-V Subgroup scope.

pattern VK_SCOPE_QUEUE_FAMILY_NV :: VkScopeNV

VK_SCOPE_QUEUE_FAMILY_NV corresponds to SPIR-V QueueFamilyKHR scope.

type FN_vkGetPhysicalDeviceCooperativeMatrixPropertiesNV = ("physicalDevice" ::: VkPhysicalDevice) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkCooperativeMatrixPropertiesNV) -> IO VkResult

type PFN_vkGetPhysicalDeviceCooperativeMatrixPropertiesNV = FunPtr FN_vkGetPhysicalDeviceCooperativeMatrixPropertiesNV

vkGetPhysicalDeviceCooperativeMatrixPropertiesNV :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("pPropertyCount" ::: Ptr Word32) -> ("pProperties" ::: Ptr VkCooperativeMatrixPropertiesNV) -> IO VkResult

vkGetPhysicalDeviceCooperativeMatrixPropertiesNV - Returns properties describing what cooperative matrix types are supported

Parameters

physicalDevice is the physical device.

pPropertyCount is a pointer to an integer related to the number of cooperative matrix properties available or queried.

pProperties is either NULL or a pointer to an array of VkCooperativeMatrixPropertiesNV structures.

Description

If pProperties is NULL, then the number of cooperative matrix properties available is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If pPropertyCount is less than the number of cooperative matrix properties available, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of cooperative matrix properties available, VK_INCOMPLETE will be returned instead of VK_SUCCESS, to indicate that not all the available cooperative matrix properties were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkCooperativeMatrixPropertiesNV structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkCooperativeMatrixPropertiesNV, VkPhysicalDevice

pattern VK_NV_COOPERATIVE_MATRIX_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_NV_COOPERATIVE_MATRIX_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_COOPERATIVE_MATRIX_PROPERTIES_NV :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_COOPERATIVE_MATRIX_FEATURES_NV :: VkStructureType

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_COOPERATIVE_MATRIX_PROPERTIES_NV :: VkStructureType