Graphics.Vulkan.C.Extensions.VK_NV_corner_sampled_image

Documentation

data VkPhysicalDeviceCornerSampledImageFeaturesNV

VkPhysicalDeviceCornerSampledImageFeaturesNV - Structure describing corner sampled image features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceCornerSampledImageFeaturesNV structure describe the following features:

Description

If the VkPhysicalDeviceCornerSampledImageFeaturesNV structure is included in the pNext chain of VkPhysicalDeviceFeatures2KHR, it is filled with values indicating whether each feature is supported. VkPhysicalDeviceCornerSampledImageFeaturesNV can also be used in the pNext chain of VkDeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

VkPhysicalDeviceCornerSampledImageFeaturesNV

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_CORNER_SAMPLED_IMAGE_FEATURES_NV

vkPNext :: Ptr ()

vkCornerSampledImage :: VkBool32

cornerSampledImage specifies whether images can be created with a VkImageCreateInfo::flags containing VK_IMAGE_CREATE_CORNER_SAMPLED_BIT_NV. See Corner-Sampled Images.

pattern VK_IMAGE_CREATE_CORNER_SAMPLED_BIT_NV :: VkImageCreateFlagBits

pattern VK_NV_CORNER_SAMPLED_IMAGE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_NV_CORNER_SAMPLED_IMAGE_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_CORNER_SAMPLED_IMAGE_FEATURES_NV :: VkStructureType