Graphics.Vulkan.C.Extensions.VK_NV_dedicated_allocation

Documentation

data VkDedicatedAllocationBufferCreateInfoNV

VkDedicatedAllocationBufferCreateInfoNV - Specify that a buffer is bound to a dedicated memory resource

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

VkDedicatedAllocationBufferCreateInfoNV

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_DEDICATED_ALLOCATION_BUFFER_CREATE_INFO_NV

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkDedicatedAllocation :: VkBool32

dedicatedAllocation specifies whether the buffer will have a dedicated allocation bound to it.

data VkDedicatedAllocationImageCreateInfoNV

VkDedicatedAllocationImageCreateInfoNV - Specify that an image is bound to a dedicated memory resource

Description

Note

Using a dedicated allocation for color and depth/stencil attachments or other large images may improve performance on some devices.

Valid Usage

If dedicatedAllocation is VK_TRUE, VkImageCreateInfo::flags must not include VK_IMAGE_CREATE_SPARSE_BINDING_BIT, VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT, or VK_IMAGE_CREATE_SPARSE_ALIASED_BIT

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DEDICATED_ALLOCATION_IMAGE_CREATE_INFO_NV

See Also

VkBool32, VkStructureType

Constructors

VkDedicatedAllocationImageCreateInfoNV

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkDedicatedAllocation :: VkBool32

dedicatedAllocation specifies whether the image will have a dedicated allocation bound to it.

data VkDedicatedAllocationMemoryAllocateInfoNV

VkDedicatedAllocationMemoryAllocateInfoNV - Specify a dedicated memory allocation resource

Valid Usage

At least one of image and buffer must be VK_NULL_HANDLE

If image is not VK_NULL_HANDLE, the image must have been created with VkDedicatedAllocationImageCreateInfoNV::dedicatedAllocation equal to VK_TRUE

If buffer is not VK_NULL_HANDLE, the buffer must have been created with VkDedicatedAllocationBufferCreateInfoNV::dedicatedAllocation equal to VK_TRUE

If image is not VK_NULL_HANDLE, VkMemoryAllocateInfo::allocationSize must equal the VkMemoryRequirements::size of the image

If buffer is not VK_NULL_HANDLE, VkMemoryAllocateInfo::allocationSize must equal the VkMemoryRequirements::size of the buffer

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DEDICATED_ALLOCATION_MEMORY_ALLOCATE_INFO_NV

If image is not VK_NULL_HANDLE, image must be a valid VkImage handle

If buffer is not VK_NULL_HANDLE, buffer must be a valid VkBuffer handle

Both of buffer, and image that are valid handles must have been created, allocated, or retrieved from the same VkDevice

See Also

VkBuffer, VkImage, VkStructureType

Constructors

VkDedicatedAllocationMemoryAllocateInfoNV

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkImage :: VkImage

image is VK_NULL_HANDLE or a handle of an image which this memory will be bound to.

vkBuffer :: VkBuffer

buffer is VK_NULL_HANDLE or a handle of a buffer which this memory will be bound to.

pattern VK_NV_DEDICATED_ALLOCATION_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_NV_DEDICATED_ALLOCATION_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_DEDICATED_ALLOCATION_BUFFER_CREATE_INFO_NV :: VkStructureType

pattern VK_STRUCTURE_TYPE_DEDICATED_ALLOCATION_IMAGE_CREATE_INFO_NV :: VkStructureType

pattern VK_STRUCTURE_TYPE_DEDICATED_ALLOCATION_MEMORY_ALLOCATE_INFO_NV :: VkStructureType