Graphics.Vulkan.C.Extensions.VK_NV_device_diagnostic_checkpoints

Documentation

data VkCheckpointDataNV

VkCheckpointDataNV - return structure for command buffer checkpoint data

Valid Usage (Implicit)

Note that the stages at which a checkpoint marker can be executed are implementation-defined and can be queried by calling vkGetPhysicalDeviceQueueFamilyProperties2.

See Also

VkPipelineStageFlagBits, VkStructureType, vkGetQueueCheckpointDataNV

Constructors

VkCheckpointDataNV

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_CHECKPOINT_DATA_NV

vkPNext :: Ptr ()

pNext must be NULL

vkStage :: VkPipelineStageFlagBits

stage indicates which pipeline stage the checkpoint marker data refers to.

vkPCheckpointMarker :: Ptr ()

pCheckpointMarker contains the value of the last checkpoint marker executed in the stage that stage refers to.

data VkQueueFamilyCheckpointPropertiesNV

VkQueueFamilyCheckpointPropertiesNV - return structure for queue family checkpoint info query

Valid Usage (Implicit)

See Also

VkPipelineStageFlags, VkStructureType

Constructors

VkQueueFamilyCheckpointPropertiesNV

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_QUEUE_FAMILY_CHECKPOINT_PROPERTIES_NV

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkCheckpointExecutionStageMask :: VkPipelineStageFlags

checkpointExecutionStageMask is a mask indicating which pipeline stages the implementation can execute checkpoint markers in.

type FN_vkCmdSetCheckpointNV = ("commandBuffer" ::: VkCommandBuffer) -> ("pCheckpointMarker" ::: Ptr ()) -> IO ()

type PFN_vkCmdSetCheckpointNV = FunPtr FN_vkCmdSetCheckpointNV

vkCmdSetCheckpointNV :: DeviceCmds -> ("commandBuffer" ::: VkCommandBuffer) -> ("pCheckpointMarker" ::: Ptr ()) -> IO ()

vkCmdSetCheckpointNV - insert diagnostic checkpoint in command stream

Parameters

commandBuffer is the command buffer that will receive the marker

pCheckpointMarker is an opaque application-provided value that will be associated with the checkpoint.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, compute, or transfer operations

Host Synchronization

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
|                 |                 | Transfer        |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer

type FN_vkGetQueueCheckpointDataNV = ("queue" ::: VkQueue) -> ("pCheckpointDataCount" ::: Ptr Word32) -> ("pCheckpointData" ::: Ptr VkCheckpointDataNV) -> IO ()

type PFN_vkGetQueueCheckpointDataNV = FunPtr FN_vkGetQueueCheckpointDataNV

vkGetQueueCheckpointDataNV :: DeviceCmds -> ("queue" ::: VkQueue) -> ("pCheckpointDataCount" ::: Ptr Word32) -> ("pCheckpointData" ::: Ptr VkCheckpointDataNV) -> IO ()

vkGetQueueCheckpointDataNV - retrieve diagnostic checkpoint data

Parameters

queue is the VkQueue object the caller would like to retrieve checkpoint data for

pCheckpointDataCount is a pointer to an integer related to the number of checkpoint markers available or queried, as described below.

pCheckpointData is either NULL or a pointer to an array of VkCheckpointDataNV structures.

Description

If pCheckpointData is NULL, then the number of checkpoint markers available is returned in pCheckpointDataCount.

Otherwise, pCheckpointDataCount must point to a variable set by the user to the number of elements in the pCheckpointData array, and on return the variable is overwritten with the number of structures actually written to pCheckpointData.

If pCheckpointDataCount is less than the number of checkpoint markers available, at most pCheckpointDataCount structures will be written.

Valid Usage

The device that queue belongs to must be in the lost state

Valid Usage (Implicit)

queue must be a valid VkQueue handle

pCheckpointDataCount must be a valid pointer to a uint32_t value

If the value referenced by pCheckpointDataCount is not 0, and pCheckpointData is not NULL, pCheckpointData must be a valid pointer to an array of pCheckpointDataCount VkCheckpointDataNV structures

See Also

VkCheckpointDataNV, VkQueue

pattern VK_NV_DEVICE_DIAGNOSTIC_CHECKPOINTS_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_NV_DEVICE_DIAGNOSTIC_CHECKPOINTS_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_CHECKPOINT_DATA_NV :: VkStructureType

pattern VK_STRUCTURE_TYPE_QUEUE_FAMILY_CHECKPOINT_PROPERTIES_NV :: VkStructureType