Graphics.Vulkan.C.Extensions.VK_NV_external_memory_capabilities

Documentation

data VkExternalImageFormatPropertiesNV

VkExternalImageFormatPropertiesNV - Structure specifying external image format properties

See Also

VkExternalMemoryFeatureFlagsNV, VkExternalMemoryHandleTypeFlagsNV, VkImageFormatProperties, vkGetPhysicalDeviceExternalImageFormatPropertiesNV

Constructors

VkExternalImageFormatPropertiesNV

Fields

vkImageFormatProperties :: VkImageFormatProperties

imageFormatProperties will be filled in as when calling vkGetPhysicalDeviceImageFormatProperties, but the values returned may vary depending on the external handle type requested.

vkExternalMemoryFeatures :: VkExternalMemoryFeatureFlagsNV

externalMemoryFeatures is a bitmask of VkExternalMemoryFeatureFlagBitsNV, indicating properties of the external memory handle type (vkGetPhysicalDeviceExternalImageFormatPropertiesNV::externalHandleType) being queried, or 0 if the external memory handle type is 0.

vkExportFromImportedHandleTypes :: VkExternalMemoryHandleTypeFlagsNV

exportFromImportedHandleTypes is a bitmask of VkExternalMemoryHandleTypeFlagBitsNV containing a bit set for every external handle type that may be used to create memory from which the handles of the type specified in vkGetPhysicalDeviceExternalImageFormatPropertiesNV::externalHandleType can be exported, or 0 if the external memory handle type is 0.

vkCompatibleHandleTypes :: VkExternalMemoryHandleTypeFlagsNV

compatibleHandleTypes is a bitmask of VkExternalMemoryHandleTypeFlagBitsNV containing a bit set for every external handle type that may be specified simultaneously with the handle type specified by vkGetPhysicalDeviceExternalImageFormatPropertiesNV::externalHandleType when calling vkAllocateMemory, or 0 if the external memory handle type is 0. compatibleHandleTypes will always contain vkGetPhysicalDeviceExternalImageFormatPropertiesNV::externalHandleType

newtype VkExternalMemoryFeatureFlagBitsNV

VkExternalMemoryFeatureFlagBitsNV - Bitmask specifying external memory features

See Also

VkExternalImageFormatPropertiesNV, VkExternalMemoryFeatureFlagsNV, vkGetPhysicalDeviceExternalImageFormatPropertiesNV

Constructors

VkExternalMemoryFeatureFlagBitsNV VkFlags

pattern VK_EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT_NV :: VkExternalMemoryFeatureFlagBitsNV

VK_EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT_NV specifies that external memory of the specified type must be created as a dedicated allocation when used in the manner specified.

pattern VK_EXTERNAL_MEMORY_FEATURE_EXPORTABLE_BIT_NV :: VkExternalMemoryFeatureFlagBitsNV

VK_EXTERNAL_MEMORY_FEATURE_EXPORTABLE_BIT_NV specifies that the implementation supports exporting handles of the specified type.

pattern VK_EXTERNAL_MEMORY_FEATURE_IMPORTABLE_BIT_NV :: VkExternalMemoryFeatureFlagBitsNV

VK_EXTERNAL_MEMORY_FEATURE_IMPORTABLE_BIT_NV specifies that the implementation supports importing handles of the specified type.

type VkExternalMemoryFeatureFlagsNV = VkExternalMemoryFeatureFlagBitsNV

VkExternalMemoryFeatureFlagsNV - Bitmask of VkExternalMemoryFeatureFlagBitsNV

Description

VkExternalMemoryFeatureFlagsNV is a bitmask type for setting a mask of zero or more VkExternalMemoryFeatureFlagBitsNV.

See Also

VkExternalImageFormatPropertiesNV, VkExternalMemoryFeatureFlagBitsNV

newtype VkExternalMemoryHandleTypeFlagBitsNV

VkExternalMemoryHandleTypeFlagBitsNV - Bitmask specifying external memory handle types

See Also

VkExternalMemoryHandleTypeFlagsNV

Constructors

VkExternalMemoryHandleTypeFlagBitsNV VkFlags

pattern VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT_NV :: VkExternalMemoryHandleTypeFlagBitsNV

VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT_NV specifies a handle to memory returned by vkGetMemoryWin32HandleNV, or one duplicated from such a handle using DuplicateHandle().

pattern VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_NV :: VkExternalMemoryHandleTypeFlagBitsNV

VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_NV specifies a handle to memory returned by vkGetMemoryWin32HandleNV.

pattern VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_BIT_NV :: VkExternalMemoryHandleTypeFlagBitsNV

VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_BIT_NV specifies a valid NT handle to memory returned by IDXGIResource1::CreateSharedHandle, or a handle duplicated from such a handle using DuplicateHandle().

pattern VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_KMT_BIT_NV :: VkExternalMemoryHandleTypeFlagBitsNV

VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_KMT_BIT_NV specifies a handle to memory returned by IDXGIResource::GetSharedHandle().

type VkExternalMemoryHandleTypeFlagsNV = VkExternalMemoryHandleTypeFlagBitsNV

VkExternalMemoryHandleTypeFlagsNV - Bitmask of VkExternalMemoryHandleTypeFlagBitsNV

Description

VkExternalMemoryHandleTypeFlagsNV is a bitmask type for setting a mask of zero or more VkExternalMemoryHandleTypeFlagBitsNV.

See Also

VkExportMemoryAllocateInfoNV, VkExternalImageFormatPropertiesNV, VkExternalMemoryHandleTypeFlagBitsNV, VkExternalMemoryImageCreateInfoNV, VkImportMemoryWin32HandleInfoNV, vkGetMemoryWin32HandleNV, vkGetPhysicalDeviceExternalImageFormatPropertiesNV

type FN_vkGetPhysicalDeviceExternalImageFormatPropertiesNV = ("physicalDevice" ::: VkPhysicalDevice) -> ("format" ::: VkFormat) -> ("type" ::: VkImageType) -> ("tiling" ::: VkImageTiling) -> ("usage" ::: VkImageUsageFlags) -> ("flags" ::: VkImageCreateFlags) -> ("externalHandleType" ::: VkExternalMemoryHandleTypeFlagsNV) -> ("pExternalImageFormatProperties" ::: Ptr VkExternalImageFormatPropertiesNV) -> IO VkResult

type PFN_vkGetPhysicalDeviceExternalImageFormatPropertiesNV = FunPtr FN_vkGetPhysicalDeviceExternalImageFormatPropertiesNV

vkGetPhysicalDeviceExternalImageFormatPropertiesNV :: InstanceCmds -> ("physicalDevice" ::: VkPhysicalDevice) -> ("format" ::: VkFormat) -> ("type" ::: VkImageType) -> ("tiling" ::: VkImageTiling) -> ("usage" ::: VkImageUsageFlags) -> ("flags" ::: VkImageCreateFlags) -> ("externalHandleType" ::: VkExternalMemoryHandleTypeFlagsNV) -> ("pExternalImageFormatProperties" ::: Ptr VkExternalImageFormatPropertiesNV) -> IO VkResult

vkGetPhysicalDeviceExternalImageFormatPropertiesNV - determine image capabilities compatible with external memory handle types

Parameters

physicalDevice is the physical device from which to query the image capabilities

format is the image format, corresponding to VkImageCreateInfo::format.

type is the image type, corresponding to VkImageCreateInfo::imageType.

tiling is the image tiling, corresponding to VkImageCreateInfo::tiling.

usage is the intended usage of the image, corresponding to VkImageCreateInfo::usage.

flags is a bitmask describing additional parameters of the image, corresponding to VkImageCreateInfo::flags.

externalHandleType is either one of the bits from VkExternalMemoryHandleTypeFlagBitsNV, or 0.

pExternalImageFormatProperties points to an instance of the VkExternalImageFormatPropertiesNV structure in which capabilities are returned.

Description

If externalHandleType is 0, pExternalImageFormatProperties::imageFormatProperties will return the same values as a call to vkGetPhysicalDeviceImageFormatProperties, and the other members of pExternalImageFormatProperties will all be 0. Otherwise, they are filled in as described for VkExternalImageFormatPropertiesNV.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_FORMAT_NOT_SUPPORTED

See Also

VkExternalImageFormatPropertiesNV, VkExternalMemoryHandleTypeFlagsNV, VkFormat, VkImageCreateFlags, VkImageTiling, VkImageType, VkImageUsageFlags, VkPhysicalDevice

pattern VK_NV_EXTERNAL_MEMORY_CAPABILITIES_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_NV_EXTERNAL_MEMORY_CAPABILITIES_SPEC_VERSION :: Integral a => a