Graphics.Vulkan.C.Extensions.VK_NV_external_memory_win32

Documentation

type DWORD = Word32

type HANDLE = Ptr ()

data SECURITY_ATTRIBUTES

Opaque data

data VkExportMemoryWin32HandleInfoNV

VkExportMemoryWin32HandleInfoNV - specify security attributes and access rights for Win32 memory handles

Description

If this structure is not present, or if pAttributes is set to NULL, default security descriptor values will be used, and child processes created by the application will not inherit the handle, as described in the MSDN documentation for “Synchronization Object Security and Access Rights”[1]. Further, if the structure is not present, the access rights will be

DXGI_SHARED_RESOURCE_READ | DXGI_SHARED_RESOURCE_WRITE

1

https://msdn.microsoft.com/en-us/library/windows/desktop/ms686670.aspx

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_EXPORT_MEMORY_WIN32_HANDLE_INFO_NV

If pAttributes is not NULL, pAttributes must be a valid pointer to a valid SECURITY_ATTRIBUTES value

See Also

VkStructureType

Constructors

VkExportMemoryWin32HandleInfoNV

Fields

vkSType :: VkStructureType

sType is the type of this structure.

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkPAttributes :: Ptr SECURITY_ATTRIBUTES

pAttributes is a pointer to a Windows SECURITY_ATTRIBUTES structure specifying security attributes of the handle.

vkDwAccess :: DWORD

dwAccess is a DWORD specifying access rights of the handle.

data VkImportMemoryWin32HandleInfoNV

VkImportMemoryWin32HandleInfoNV - import Win32 memory created on the same physical device

Description

If handleType is 0, this structure is ignored by consumers of the VkMemoryAllocateInfo structure it is chained from.

Valid Usage (Implicit)

See Also

VkExternalMemoryHandleTypeFlagsNV, VkStructureType

Constructors

VkImportMemoryWin32HandleInfoNV

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_IMPORT_MEMORY_WIN32_HANDLE_INFO_NV

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkHandleType :: VkExternalMemoryHandleTypeFlagsNV

handleType must be a valid combination of VkExternalMemoryHandleTypeFlagBitsNV values

vkHandle :: HANDLE

handle must be a valid handle to memory, obtained as specified by handleType.

type FN_vkGetMemoryWin32HandleNV = ("device" ::: VkDevice) -> ("memory" ::: VkDeviceMemory) -> ("handleType" ::: VkExternalMemoryHandleTypeFlagsNV) -> ("pHandle" ::: Ptr HANDLE) -> IO VkResult

type PFN_vkGetMemoryWin32HandleNV = FunPtr FN_vkGetMemoryWin32HandleNV

vkGetMemoryWin32HandleNV :: DeviceCmds -> ("device" ::: VkDevice) -> ("memory" ::: VkDeviceMemory) -> ("handleType" ::: VkExternalMemoryHandleTypeFlagsNV) -> ("pHandle" ::: Ptr HANDLE) -> IO VkResult

vkGetMemoryWin32HandleNV - retrieve Win32 handle to a device memory object

Parameters

device is the logical device that owns the memory.

memory is the VkDeviceMemory object.

handleType is a bitmask of VkExternalMemoryHandleTypeFlagBitsNV containing a single bit specifying the type of handle requested.

handle points to a Windows HANDLE in which the handle is returned.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_TOO_MANY_OBJECTS

VK_ERROR_OUT_OF_HOST_MEMORY

See Also

VkDevice, VkDeviceMemory, VkExternalMemoryHandleTypeFlagsNV

pattern VK_NV_EXTERNAL_MEMORY_WIN32_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_NV_EXTERNAL_MEMORY_WIN32_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_EXPORT_MEMORY_WIN32_HANDLE_INFO_NV :: VkStructureType

pattern VK_STRUCTURE_TYPE_IMPORT_MEMORY_WIN32_HANDLE_INFO_NV :: VkStructureType