Graphics.Vulkan.C.Extensions.VK_NV_fragment_coverage_to_color

Documentation

newtype VkPipelineCoverageToColorStateCreateFlagsNV

VkPipelineCoverageToColorStateCreateFlagsNV - Reserved for future use

Description

VkPipelineCoverageToColorStateCreateFlagsNV is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkPipelineCoverageToColorStateCreateInfoNV

Constructors

VkPipelineCoverageToColorStateCreateFlagsNV VkFlags

data VkPipelineCoverageToColorStateCreateInfoNV

VkPipelineCoverageToColorStateCreateInfoNV - Structure specifying whether fragment coverage replaces a color

Description

If coverageToColorEnable is VK_TRUE, the fragment coverage information is treated as a bitmask with one bit for each sample (as in the Sample Mask section), and this bitmask replaces the first component of the color value corresponding to the fragment shader output location with Location equal to coverageToColorLocation and Index equal to zero. If the color attachment format has fewer bits than the sample coverage, the low bits of the sample coverage bitmask are taken without any clamping. If the color attachment format has more bits than the sample coverage, the high bits of the sample coverage bitmask are filled with zeros.

If Sample Shading is in use, the coverage bitmask only has bits set for samples that correspond to the fragment shader invocation that shades those samples.

This pipeline stage occurs after sample counting and before blending, and is always performed after fragment shading regardless of the setting of EarlyFragmentTests.

If coverageToColorEnable is VK_FALSE, these operations are skipped. If this structure is not present, it is as if coverageToColorEnable is VK_FALSE.

Valid Usage

If coverageToColorEnable is VK_TRUE, then the render pass subpass indicated by VkGraphicsPipelineCreateInfo::renderPass and VkGraphicsPipelineCreateInfo::subpass must have a color attachment at the location selected by coverageToColorLocation, with a VkFormat of VK_FORMAT_R8_UINT, VK_FORMAT_R8_SINT, VK_FORMAT_R16_UINT, VK_FORMAT_R16_SINT, VK_FORMAT_R32_UINT, or VK_FORMAT_R32_SINT

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PIPELINE_COVERAGE_TO_COLOR_STATE_CREATE_INFO_NV

flags must be 0

See Also

VkBool32, VkPipelineCoverageToColorStateCreateFlagsNV, VkStructureType

Constructors

VkPipelineCoverageToColorStateCreateInfoNV

Fields

vkSType :: VkStructureType

sType is the type of this structure

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure

vkFlags :: VkPipelineCoverageToColorStateCreateFlagsNV

flags is reserved for future use.

vkCoverageToColorEnable :: VkBool32

coverageToColorEnable controls whether the fragment coverage value replaces a fragment color output.

vkCoverageToColorLocation :: Word32

coverageToColorLocation controls which fragment shader color output value is replaced.

pattern VK_NV_FRAGMENT_COVERAGE_TO_COLOR_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_NV_FRAGMENT_COVERAGE_TO_COLOR_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_PIPELINE_COVERAGE_TO_COLOR_STATE_CREATE_INFO_NV :: VkStructureType