Graphics.Vulkan.C.Extensions.VK_NV_viewport_swizzle

Documentation

newtype VkPipelineViewportSwizzleStateCreateFlagsNV

VkPipelineViewportSwizzleStateCreateFlagsNV - Reserved for future use

Description

VkPipelineViewportSwizzleStateCreateFlagsNV is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkPipelineViewportSwizzleStateCreateInfoNV

Constructors

VkPipelineViewportSwizzleStateCreateFlagsNV VkFlags

data VkPipelineViewportSwizzleStateCreateInfoNV

VkPipelineViewportSwizzleStateCreateInfoNV - Structure specifying swizzle applied to primitive clip coordinates

Valid Usage (Implicit)

See Also

VkPipelineViewportSwizzleStateCreateFlagsNV, VkStructureType, VkViewportSwizzleNV

Constructors

VkPipelineViewportSwizzleStateCreateInfoNV

Fields

vkSType :: VkStructureType

sType must be VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_SWIZZLE_STATE_CREATE_INFO_NV

vkPNext :: Ptr ()

pNext is NULL or a pointer to an extension-specific structure.

vkFlags :: VkPipelineViewportSwizzleStateCreateFlagsNV

flags must be 0

vkViewportCount :: Word32

viewportCount must be greater than 0

vkPViewportSwizzles :: Ptr VkViewportSwizzleNV

pViewportSwizzles must be a valid pointer to an array of viewportCount valid VkViewportSwizzleNV structures

newtype VkViewportCoordinateSwizzleNV

VkViewportCoordinateSwizzleNV - Specify how a viewport coordinate is swizzled

Description

These values are described in detail in Viewport Swizzle.

See Also

VkViewportSwizzleNV

Constructors

VkViewportCoordinateSwizzleNV Int32

pattern VK_VIEWPORT_COORDINATE_SWIZZLE_POSITIVE_X_NV :: VkViewportCoordinateSwizzleNV

pattern VK_VIEWPORT_COORDINATE_SWIZZLE_NEGATIVE_X_NV :: VkViewportCoordinateSwizzleNV

pattern VK_VIEWPORT_COORDINATE_SWIZZLE_POSITIVE_Y_NV :: VkViewportCoordinateSwizzleNV

pattern VK_VIEWPORT_COORDINATE_SWIZZLE_NEGATIVE_Y_NV :: VkViewportCoordinateSwizzleNV

pattern VK_VIEWPORT_COORDINATE_SWIZZLE_POSITIVE_Z_NV :: VkViewportCoordinateSwizzleNV

pattern VK_VIEWPORT_COORDINATE_SWIZZLE_NEGATIVE_Z_NV :: VkViewportCoordinateSwizzleNV

pattern VK_VIEWPORT_COORDINATE_SWIZZLE_POSITIVE_W_NV :: VkViewportCoordinateSwizzleNV

pattern VK_VIEWPORT_COORDINATE_SWIZZLE_NEGATIVE_W_NV :: VkViewportCoordinateSwizzleNV

data VkViewportSwizzleNV

VkViewportSwizzleNV - Structure specifying a viewport swizzle

Valid Usage (Implicit)

See Also

VkPipelineViewportSwizzleStateCreateInfoNV, VkViewportCoordinateSwizzleNV

Constructors

VkViewportSwizzleNV

Fields

vkX :: VkViewportCoordinateSwizzleNV

x must be a valid VkViewportCoordinateSwizzleNV value

vkY :: VkViewportCoordinateSwizzleNV

y must be a valid VkViewportCoordinateSwizzleNV value

vkZ :: VkViewportCoordinateSwizzleNV

z must be a valid VkViewportCoordinateSwizzleNV value

vkW :: VkViewportCoordinateSwizzleNV

w must be a valid VkViewportCoordinateSwizzleNV value

pattern VK_NV_VIEWPORT_SWIZZLE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern VK_NV_VIEWPORT_SWIZZLE_SPEC_VERSION :: Integral a => a

pattern VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_SWIZZLE_STATE_CREATE_INFO_NV :: VkStructureType