Graphics.Vulkan.Core10.AllocationCallbacks

Documentation

data AllocationCallbacks

VkAllocationCallbacks - Structure containing callback function pointers for memory allocation

Valid Usage

pfnAllocation must be a valid pointer to a valid user-defined PFN_vkAllocationFunction

pfnReallocation must be a valid pointer to a valid user-defined PFN_vkReallocationFunction

pfnFree must be a valid pointer to a valid user-defined PFN_vkFreeFunction

If either of pfnInternalAllocation or pfnInternalFree is not NULL, both must be valid callbacks

See Also

PFN_vkAllocationFunction, PFN_vkFreeFunction, PFN_vkInternalAllocationNotification, PFN_vkInternalFreeNotification, PFN_vkReallocationFunction, allocateMemory, createAccelerationStructureNV, createAndroidSurfaceKHR, createBuffer, createBufferView, createCommandPool, createComputePipelines, createDebugReportCallbackEXT, createDebugUtilsMessengerEXT, createDescriptorPool, createDescriptorSetLayout, createDescriptorUpdateTemplate, createDescriptorUpdateTemplateKHR, createDevice, createDisplayModeKHR, createDisplayPlaneSurfaceKHR, createEvent, createFence, createFramebuffer, createGraphicsPipelines, createHeadlessSurfaceEXT, createIOSSurfaceMVK, createImage, createImagePipeSurfaceFUCHSIA, createImageView, createIndirectCommandsLayoutNVX, createInstance, createMacOSSurfaceMVK, createMetalSurfaceEXT, createObjectTableNVX, createPipelineCache, createPipelineLayout, createQueryPool, createRayTracingPipelinesNV, createRenderPass, createRenderPass2, createRenderPass2KHR, createSampler, createSamplerYcbcrConversion, createSamplerYcbcrConversionKHR, createSemaphore, createShaderModule, createSharedSwapchainsKHR, createStreamDescriptorSurfaceGGP, createSwapchainKHR, createValidationCacheEXT, createViSurfaceNN, createWaylandSurfaceKHR, createWin32SurfaceKHR, createXcbSurfaceKHR, createXlibSurfaceKHR, destroyAccelerationStructureNV, destroyBuffer, destroyBufferView, destroyCommandPool, destroyDebugReportCallbackEXT, destroyDebugUtilsMessengerEXT, destroyDescriptorPool, destroyDescriptorSetLayout, destroyDescriptorUpdateTemplate, destroyDescriptorUpdateTemplateKHR, destroyDevice, destroyEvent, destroyFence, destroyFramebuffer, destroyImage, destroyImageView, destroyIndirectCommandsLayoutNVX, destroyInstance, destroyObjectTableNVX, destroyPipeline, destroyPipelineCache, destroyPipelineLayout, destroyQueryPool, destroyRenderPass, destroySampler, destroySamplerYcbcrConversion, destroySamplerYcbcrConversionKHR, destroySemaphore, destroyShaderModule, destroySurfaceKHR, destroySwapchainKHR, destroyValidationCacheEXT, freeMemory, registerDeviceEventEXT, registerDisplayEventEXT

Constructors

AllocationCallbacks

Fields

userData :: Ptr ()

pUserData is a value to be interpreted by the implementation of the callbacks. When any of the callbacks in AllocationCallbacks are called, the Vulkan implementation will pass this value as the first parameter to the callback. This value can vary each time an allocator is passed into a command, even when the same object takes an allocator in multiple commands.

pfnAllocation :: PFN_vkAllocationFunction

pfnAllocation is a PFN_vkAllocationFunction pointer to an application-defined memory allocation function.

pfnReallocation :: PFN_vkReallocationFunction

pfnReallocation is a PFN_vkReallocationFunction pointer to an application-defined memory reallocation function.

pfnFree :: PFN_vkFreeFunction

pfnFree is a PFN_vkFreeFunction pointer to an application-defined memory free function.

pfnInternalAllocation :: PFN_vkInternalAllocationNotification

pfnInternalAllocation is a PFN_vkInternalAllocationNotification pointer to an application-defined function that is called by the implementation when the implementation makes internal allocations.

pfnInternalFree :: PFN_vkInternalFreeNotification

pfnInternalFree is a PFN_vkInternalFreeNotification pointer to an application-defined function that is called by the implementation when the implementation frees internal allocations.