Graphics.Vulkan.Core10.BaseType

Documentation

boolToBool32 :: Bool -> Bool32

bool32ToBool :: Bool32 -> Bool

newtype Bool32

VkBool32 - Vulkan boolean type

Description

TRUE represents a boolean True (integer 1) value, and FALSE a boolean False (integer 0) value.

All values returned from a Vulkan implementation in a Bool32 will be either TRUE or FALSE.

Applications must not pass any other values than TRUE or FALSE into a Vulkan implementation where a Bool32 is expected.

See Also

CommandBufferInheritanceConditionalRenderingInfoEXT, CommandBufferInheritanceInfo, DedicatedAllocationBufferCreateInfoNV, DedicatedAllocationImageCreateInfoNV, DescriptorSetLayoutSupport, DeviceGeneratedCommandsFeaturesNVX, DisplayNativeHdrSurfaceCapabilitiesAMD, DisplayPresentInfoKHR, DisplayPropertiesKHR, FilterCubicImageViewImageFormatPropertiesEXT, MemoryDedicatedRequirements, PerformanceOverrideInfoINTEL, PerformanceValueDataINTEL, PhysicalDevice16BitStorageFeatures, PhysicalDevice8BitStorageFeatures, PhysicalDeviceASTCDecodeFeaturesEXT, PhysicalDeviceBlendOperationAdvancedFeaturesEXT, PhysicalDeviceBlendOperationAdvancedPropertiesEXT, PhysicalDeviceBufferDeviceAddressFeatures, PhysicalDeviceBufferDeviceAddressFeaturesEXT, PhysicalDeviceCoherentMemoryFeaturesAMD, PhysicalDeviceComputeShaderDerivativesFeaturesNV, PhysicalDeviceConditionalRenderingFeaturesEXT, PhysicalDeviceConservativeRasterizationPropertiesEXT, PhysicalDeviceCooperativeMatrixFeaturesNV, PhysicalDeviceCornerSampledImageFeaturesNV, PhysicalDeviceCoverageReductionModeFeaturesNV, PhysicalDeviceDedicatedAllocationImageAliasingFeaturesNV, PhysicalDeviceDepthClipEnableFeaturesEXT, PhysicalDeviceDepthStencilResolveProperties, PhysicalDeviceDescriptorIndexingFeatures, PhysicalDeviceDescriptorIndexingProperties, PhysicalDeviceExclusiveScissorFeaturesNV, PhysicalDeviceFeatures, PhysicalDeviceFloatControlsProperties, PhysicalDeviceFragmentDensityMapFeaturesEXT, PhysicalDeviceFragmentDensityMapPropertiesEXT, PhysicalDeviceFragmentShaderBarycentricFeaturesNV, PhysicalDeviceFragmentShaderInterlockFeaturesEXT, PhysicalDeviceGroupProperties, PhysicalDeviceHostQueryResetFeatures, PhysicalDeviceIDProperties, PhysicalDeviceImagelessFramebufferFeatures, PhysicalDeviceIndexTypeUint8FeaturesEXT, PhysicalDeviceInlineUniformBlockFeaturesEXT, PhysicalDeviceLimits, PhysicalDeviceLineRasterizationFeaturesEXT, PhysicalDeviceMemoryPriorityFeaturesEXT, PhysicalDeviceMeshShaderFeaturesNV, PhysicalDeviceMultiviewFeatures, PhysicalDeviceMultiviewPerViewAttributesPropertiesNVX, PhysicalDevicePerformanceQueryFeaturesKHR, PhysicalDevicePerformanceQueryPropertiesKHR, PhysicalDevicePipelineExecutablePropertiesFeaturesKHR, PhysicalDeviceProtectedMemoryFeatures, PhysicalDeviceProtectedMemoryProperties, PhysicalDeviceRepresentativeFragmentTestFeaturesNV, PhysicalDeviceSampleLocationsPropertiesEXT, PhysicalDeviceSamplerFilterMinmaxProperties, PhysicalDeviceSamplerYcbcrConversionFeatures, PhysicalDeviceScalarBlockLayoutFeatures, PhysicalDeviceSeparateDepthStencilLayoutsFeatures, PhysicalDeviceShaderAtomicInt64Features, PhysicalDeviceShaderClockFeaturesKHR, PhysicalDeviceShaderDemoteToHelperInvocationFeaturesEXT, PhysicalDeviceShaderDrawParametersFeatures, PhysicalDeviceShaderFloat16Int8Features, PhysicalDeviceShaderImageFootprintFeaturesNV, PhysicalDeviceShaderIntegerFunctions2FeaturesINTEL, PhysicalDeviceShaderSMBuiltinsFeaturesNV, PhysicalDeviceShaderSubgroupExtendedTypesFeatures, PhysicalDeviceShadingRateImageFeaturesNV, PhysicalDeviceSparseProperties, PhysicalDeviceSubgroupProperties, PhysicalDeviceSubgroupSizeControlFeaturesEXT, PhysicalDeviceTexelBufferAlignmentFeaturesEXT, PhysicalDeviceTexelBufferAlignmentPropertiesEXT, PhysicalDeviceTextureCompressionASTCHDRFeaturesEXT, PhysicalDeviceTimelineSemaphoreFeatures, PhysicalDeviceTransformFeedbackFeaturesEXT, PhysicalDeviceTransformFeedbackPropertiesEXT, PhysicalDeviceUniformBufferStandardLayoutFeatures, PhysicalDeviceVariablePointersFeatures, PhysicalDeviceVertexAttributeDivisorFeaturesEXT, PhysicalDeviceVulkan11Features, PhysicalDeviceVulkan11Properties, PhysicalDeviceVulkan12Features, PhysicalDeviceVulkan12Properties, PhysicalDeviceVulkanMemoryModelFeatures, PhysicalDeviceYcbcrImageArraysFeaturesEXT, PipelineColorBlendAdvancedStateCreateInfoEXT, PipelineColorBlendAttachmentState, PipelineColorBlendStateCreateInfo, PipelineCoverageModulationStateCreateInfoNV, PipelineCoverageToColorStateCreateInfoNV, PipelineDepthStencilStateCreateInfo, PipelineExecutableInternalRepresentationKHR, PipelineExecutableStatisticValueKHR, PipelineInputAssemblyStateCreateInfo, PipelineMultisampleStateCreateInfo, PipelineRasterizationDepthClipStateCreateInfoEXT, PipelineRasterizationLineStateCreateInfoEXT, PipelineRasterizationStateCreateInfo, PipelineRepresentativeFragmentTestStateCreateInfoNV, PipelineSampleLocationsStateCreateInfoEXT, PipelineViewportShadingRateImageStateCreateInfoNV, PipelineViewportWScalingStateCreateInfoNV, ProtectedSubmitInfo, SamplerCreateInfo, SamplerYcbcrConversionCreateInfo, SurfaceCapabilitiesFullScreenExclusiveEXT, SurfaceProtectedCapabilitiesKHR, SwapchainCreateInfoKHR, SwapchainDisplayNativeHdrCreateInfoAMD, TextureLODGatherFormatPropertiesAMD, cmdBuildAccelerationStructureNV, getPhysicalDeviceSurfaceSupportKHR, setLocalDimmingAMD, waitForFences

Constructors

Bool32 Int32

Bundled Patterns

pattern FALSE :: Bool32

pattern TRUE :: Bool32

type SampleMask = Word32

VkSampleMask - Mask of sample coverage information

See Also

PipelineMultisampleStateCreateInfo

type Flags = Word32

VkFlags - Vulkan bitmasks

Description

Bitmasks are passed to many commands and structures to compactly represent options, but Flags is not used directly in the API. Instead, a Vk*Flags type which is an alias of Flags, and whose name matches the corresponding Vk*FlagBits that are valid for that type, is used.

Any Vk*Flags member or parameter used in the API as an input must be a valid combination of bit flags. A valid combination is either zero or the bitwise OR of valid bit flags. A bit flag is valid if:

The bit flag is defined as part of the Vk*FlagBits type, where the bits type is obtained by taking the flag type and replacing the trailing Flags with FlagBits. For example, a flag value of type ColorComponentFlags must contain only bit flags defined by ColorComponentFlagBits.

The flag is allowed in the context in which it is being used. For example, in some cases, certain bit flags or combinations of bit flags are mutually exclusive.

Any Vk*Flags member or parameter returned from a query command or otherwise output from Vulkan to the application may contain bit flags undefined in its corresponding Vk*FlagBits type. An application cannot rely on the state of these unspecified bits.

Only the low-order 31 bits (bit positions zero through 30) are available for use as flag bits.

Note

This restriction is due to poorly defined behavior by C compilers given a C enumerant value of 0x80000000. In some cases adding this enumerant value may increase the size of the underlying Vk*FlagBits type, breaking the ABI.

See Also

ColorComponentFlags

type DeviceSize = Word64

VkDeviceSize - Vulkan device memory size and offsets

See Also

AccelerationStructureCreateInfoNV, AndroidHardwareBufferPropertiesANDROID, BindAccelerationStructureMemoryInfoNV, BindBufferMemoryInfo, BindImageMemoryInfo, BufferCopy, BufferCreateInfo, BufferImageCopy, BufferMemoryBarrier, BufferViewCreateInfo, CmdProcessCommandsInfoNVX, ConditionalRenderingBeginInfoEXT, DescriptorBufferInfo, GeometryAABBNV, GeometryTrianglesNV, ImageFormatProperties, IndirectCommandsTokenNVX, MappedMemoryRange, MemoryAllocateInfo, MemoryHeap, MemoryRequirements, PhysicalDeviceExternalMemoryHostPropertiesEXT, PhysicalDeviceLimits, PhysicalDeviceMaintenance3Properties, PhysicalDeviceMemoryBudgetPropertiesEXT, PhysicalDeviceTexelBufferAlignmentPropertiesEXT, PhysicalDeviceTransformFeedbackPropertiesEXT, PhysicalDeviceVulkan11Properties, SparseImageMemoryBind, SparseImageMemoryRequirements, SparseMemoryBind, SubresourceLayout, bindBufferMemory, bindImageMemory, cmdBeginTransformFeedbackEXT, cmdBindIndexBuffer, cmdBindTransformFeedbackBuffersEXT, cmdBindVertexBuffers, cmdBuildAccelerationStructureNV, cmdCopyQueryPoolResults, cmdDispatchIndirect, cmdDrawIndexedIndirect, cmdDrawIndexedIndirectCount, cmdDrawIndexedIndirectCountAMD, cmdDrawIndexedIndirectCountKHR, cmdDrawIndirect, cmdDrawIndirectByteCountEXT, cmdDrawIndirectCount, cmdDrawIndirectCountAMD, cmdDrawIndirectCountKHR, cmdDrawMeshTasksIndirectCountNV, cmdDrawMeshTasksIndirectNV, cmdEndTransformFeedbackEXT, cmdFillBuffer, cmdTraceRaysNV, cmdUpdateBuffer, cmdWriteBufferMarkerAMD, getDeviceMemoryCommitment, getQueryPoolResults, mapMemory

type DeviceAddress = Word64

VkDeviceAddress - Vulkan device address type

See Also

BufferDeviceAddressCreateInfoEXT