Graphics.Vulkan.Core10.Buffer

Documentation

createBuffer :: PokeChain a => Device -> BufferCreateInfo a -> ("allocator" ::: Maybe AllocationCallbacks) -> IO Buffer

vkCreateBuffer - Create a new buffer object

Parameters

Device is the logical device that creates the buffer object.

pCreateInfo is a pointer to a BufferCreateInfo structure containing parameters affecting creation of the buffer.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pBuffer is a pointer to a Buffer handle in which the resulting buffer object is returned.

Valid Usage

If the Flags member of pCreateInfo includes BUFFER_CREATE_SPARSE_BINDING_BIT, creating this Buffer must not cause the total required sparse memory for all currently valid sparse resources on the device to exceed PhysicalDeviceLimits::sparseAddressSpaceSize

Valid Usage (Implicit)

Device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid BufferCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pBuffer must be a valid pointer to a Buffer handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_INVALID_OPAQUE_CAPTURE_ADDRESS_KHR

See Also

AllocationCallbacks, Buffer, BufferCreateInfo, Device

withBuffer :: PokeChain a => Device -> BufferCreateInfo a -> Maybe AllocationCallbacks -> (Buffer -> IO r) -> IO r

A safe wrapper for createBuffer and destroyBuffer using bracket

The allocated value must not be returned from the provided computation

destroyBuffer :: Device -> Buffer -> ("allocator" ::: Maybe AllocationCallbacks) -> IO ()

vkDestroyBuffer - Destroy a buffer object

Parameters

Device is the logical device that destroys the buffer.

Buffer is the buffer to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to Buffer, either directly or via a BufferView, must have completed execution

If AllocationCallbacks were provided when Buffer was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when Buffer was created, pAllocator must be NULL

Valid Usage (Implicit)

Device must be a valid Device handle

If Buffer is not NULL_HANDLE, Buffer must be a valid Buffer handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If Buffer is a valid handle, it must have been created, allocated, or retrieved from Device

Host Synchronization

Host access to Buffer must be externally synchronized

See Also

AllocationCallbacks, Buffer, Device

data BufferCreateInfo (es :: [Type])

VkBufferCreateInfo - Structure specifying the parameters of a newly created buffer object

Valid Usage

size must be greater than 0

If SharingMode is SHARING_MODE_CONCURRENT, pQueueFamilyIndices must be a valid pointer to an array of queueFamilyIndexCount uint32_t values

If SharingMode is SHARING_MODE_CONCURRENT, queueFamilyIndexCount must be greater than 1

If SharingMode is SHARING_MODE_CONCURRENT, each element of pQueueFamilyIndices must be unique and must be less than pQueueFamilyPropertyCount returned by either getPhysicalDeviceQueueFamilyProperties or getPhysicalDeviceQueueFamilyProperties2 for the PhysicalDevice that was used to create Device

If the sparse bindings feature is not enabled, Flags must not contain BUFFER_CREATE_SPARSE_BINDING_BIT

If the sparse buffer residency feature is not enabled, Flags must not contain BUFFER_CREATE_SPARSE_RESIDENCY_BIT

If the sparse aliased residency feature is not enabled, Flags must not contain BUFFER_CREATE_SPARSE_ALIASED_BIT

If Flags contains BUFFER_CREATE_SPARSE_RESIDENCY_BIT or BUFFER_CREATE_SPARSE_ALIASED_BIT, it must also contain BUFFER_CREATE_SPARSE_BINDING_BIT

If the pNext chain includes a ExternalMemoryBufferCreateInfo structure, its handleTypes member must only contain bits that are also in ExternalBufferProperties::externalMemoryProperties.compatibleHandleTypes, as returned by getPhysicalDeviceExternalBufferProperties with pExternalBufferInfo->handleType equal to any one of the handle types specified in ExternalMemoryBufferCreateInfo::handleTypes

If the protected memory feature is not enabled, Flags must not contain BUFFER_CREATE_PROTECTED_BIT

If any of the bits BUFFER_CREATE_SPARSE_BINDING_BIT, BUFFER_CREATE_SPARSE_RESIDENCY_BIT, or BUFFER_CREATE_SPARSE_ALIASED_BIT are set, BUFFER_CREATE_PROTECTED_BIT must not also be set

If the pNext chain includes a DedicatedAllocationBufferCreateInfoNV structure, and the dedicatedAllocation member of the chained structure is TRUE, then Flags must not include BUFFER_CREATE_SPARSE_BINDING_BIT, BUFFER_CREATE_SPARSE_RESIDENCY_BIT, or BUFFER_CREATE_SPARSE_ALIASED_BIT

If 'Graphics.Vulkan.Extensions.VK_EXT_buffer_device_address.BufferDeviceAddressCreateInfoEXT'::'Graphics.Vulkan.Core10.BaseType.DeviceAddress' is not zero, Flags must include BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT

If BufferOpaqueCaptureAddressCreateInfo::opaqueCaptureAddress is not zero, Flags must include BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT

If Flags includes BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT, the bufferDeviceAddressCaptureReplay or ::bufferDeviceAddressCaptureReplay feature must be enabled

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_BUFFER_CREATE_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of BufferDeviceAddressCreateInfoEXT, BufferOpaqueCaptureAddressCreateInfo, DedicatedAllocationBufferCreateInfoNV, or ExternalMemoryBufferCreateInfo

The sType value of each struct in the pNext chain must be unique

Flags must be a valid combination of BufferCreateFlagBits values

usage must be a valid combination of BufferUsageFlagBits values

usage must not be 0

SharingMode must be a valid SharingMode value

See Also

BufferCreateFlags, BufferUsageFlags, DeviceSize, SharingMode, StructureType, createBuffer

Constructors

BufferCreateInfo

Fields

next :: Chain es

pNext is NULL or a pointer to an extension-specific structure.

flags :: BufferCreateFlags

Flags is a bitmask of BufferCreateFlagBits specifying additional parameters of the buffer.

size :: DeviceSize

size is the size in bytes of the buffer to be created.

usage :: BufferUsageFlags

usage is a bitmask of BufferUsageFlagBits specifying allowed usages of the buffer.

sharingMode :: SharingMode

SharingMode is a SharingMode value specifying the sharing mode of the buffer when it will be accessed by multiple queue families.

queueFamilyIndices :: Vector Word32

pQueueFamilyIndices is a list of queue families that will access this buffer (ignored if SharingMode is not SHARING_MODE_CONCURRENT).