Graphics.Vulkan.Core10.CommandBuffer

Documentation

withCStructCommandBufferAllocateInfo :: CommandBufferAllocateInfo -> (VkCommandBufferAllocateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkCommandBufferAllocateInfo and marshal a CommandBufferAllocateInfo into it. The VkCommandBufferAllocateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructCommandBufferAllocateInfo :: VkCommandBufferAllocateInfo -> IO CommandBufferAllocateInfo

A function to read a VkCommandBufferAllocateInfo and all additional structures in the pointer chain into a CommandBufferAllocateInfo.

data CommandBufferAllocateInfo

VkCommandBufferAllocateInfo - Structure specifying the allocation parameters for command buffer object

Valid Usage (Implicit)

See Also

VkCommandBufferLevel, VkCommandPool, VkStructureType, vkAllocateCommandBuffers

Constructors

CommandBufferAllocateInfo

Fields

next :: Maybe SomeVkStruct

commandPool :: CommandPool

level :: CommandBufferLevel

commandBufferCount :: Word32

withCStructCommandBufferBeginInfo :: CommandBufferBeginInfo -> (VkCommandBufferBeginInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkCommandBufferBeginInfo and marshal a CommandBufferBeginInfo into it. The VkCommandBufferBeginInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructCommandBufferBeginInfo :: VkCommandBufferBeginInfo -> IO CommandBufferBeginInfo

A function to read a VkCommandBufferBeginInfo and all additional structures in the pointer chain into a CommandBufferBeginInfo.

data CommandBufferBeginInfo

VkCommandBufferBeginInfo - Structure specifying a command buffer begin operation

Valid Usage

If flags contains VK_COMMAND_BUFFER_USAGE_RENDER_PASS_CONTINUE_BIT, the renderPass member of pInheritanceInfo must be a valid VkRenderPass

If flags contains VK_COMMAND_BUFFER_USAGE_RENDER_PASS_CONTINUE_BIT, the subpass member of pInheritanceInfo must be a valid subpass index within the renderPass member of pInheritanceInfo

If flags contains VK_COMMAND_BUFFER_USAGE_RENDER_PASS_CONTINUE_BIT, the framebuffer member of pInheritanceInfo must be either VK_NULL_HANDLE, or a valid VkFramebuffer that is compatible with the renderPass member of pInheritanceInfo

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO

pNext must be NULL or a pointer to a valid instance of VkDeviceGroupCommandBufferBeginInfo

flags must be a valid combination of VkCommandBufferUsageFlagBits values

See Also

VkCommandBufferInheritanceInfo, VkCommandBufferUsageFlags, VkStructureType, vkBeginCommandBuffer

Constructors

CommandBufferBeginInfo

Fields

next :: Maybe SomeVkStruct

flags :: CommandBufferUsageFlags

inheritanceInfo :: Maybe CommandBufferInheritanceInfo

withCStructCommandBufferInheritanceInfo :: CommandBufferInheritanceInfo -> (VkCommandBufferInheritanceInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkCommandBufferInheritanceInfo and marshal a CommandBufferInheritanceInfo into it. The VkCommandBufferInheritanceInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructCommandBufferInheritanceInfo :: VkCommandBufferInheritanceInfo -> IO CommandBufferInheritanceInfo

A function to read a VkCommandBufferInheritanceInfo and all additional structures in the pointer chain into a CommandBufferInheritanceInfo.

data CommandBufferInheritanceInfo

VkCommandBufferInheritanceInfo - Structure specifying command buffer inheritance info

Valid Usage

If the inherited queries feature is not enabled, occlusionQueryEnable must be VK_FALSE

If the inherited queries feature is enabled, queryFlags must be a valid combination of VkQueryControlFlagBits values

If the pipeline statistics queries feature is not enabled, pipelineStatistics must be 0

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_COMMAND_BUFFER_INHERITANCE_INFO

pNext must be NULL or a pointer to a valid instance of VkCommandBufferInheritanceConditionalRenderingInfoEXT

Both of framebuffer, and renderPass that are valid handles must have been created, allocated, or retrieved from the same VkDevice

See Also

VkBool32, VkCommandBufferBeginInfo, VkFramebuffer, VkQueryControlFlags, VkQueryPipelineStatisticFlags, VkRenderPass, VkStructureType

Constructors

CommandBufferInheritanceInfo

Fields

next :: Maybe SomeVkStruct

renderPass :: RenderPass

subpass :: Word32

framebuffer :: Framebuffer

occlusionQueryEnable :: Bool

queryFlags :: QueryControlFlags

pipelineStatistics :: QueryPipelineStatisticFlags

type CommandBufferLevel = VkCommandBufferLevel

VkCommandBufferLevel - Enumerant specifying a command buffer level

See Also

VkCommandBufferAllocateInfo

pattern COMMAND_BUFFER_LEVEL_PRIMARY :: a ~ CommandBufferLevel => a

VK_COMMAND_BUFFER_LEVEL_PRIMARY specifies a primary command buffer.

pattern COMMAND_BUFFER_LEVEL_SECONDARY :: a ~ CommandBufferLevel => a

VK_COMMAND_BUFFER_LEVEL_SECONDARY specifies a secondary command buffer.

type CommandBufferResetFlagBits = VkCommandBufferResetFlagBits

VkCommandBufferResetFlagBits - Bitmask controlling behavior of a command buffer reset

See Also

VkCommandBufferResetFlags

pattern COMMAND_BUFFER_RESET_RELEASE_RESOURCES_BIT :: a ~ CommandBufferResetFlagBits => a

VK_COMMAND_BUFFER_RESET_RELEASE_RESOURCES_BIT specifies that most or all memory resources currently owned by the command buffer should be returned to the parent command pool. If this flag is not set, then the command buffer may hold onto memory resources and reuse them when recording commands. commandBuffer is moved to the initial state.

type CommandBufferResetFlags = CommandBufferResetFlagBits

VkCommandBufferResetFlags - Bitmask of VkCommandBufferResetFlagBits

Description

VkCommandBufferResetFlags is a bitmask type for setting a mask of zero or more VkCommandBufferResetFlagBits.

See Also

VkCommandBufferResetFlagBits, vkResetCommandBuffer

type CommandBufferUsageFlagBits = VkCommandBufferUsageFlagBits

VkCommandBufferUsageFlagBits - Bitmask specifying usage behavior for command buffer

See Also

VkCommandBufferUsageFlags

pattern COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT :: a ~ CommandBufferUsageFlagBits => a

VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT specifies that each recording of the command buffer will only be submitted once, and the command buffer will be reset and recorded again between each submission.

pattern COMMAND_BUFFER_USAGE_RENDER_PASS_CONTINUE_BIT :: a ~ CommandBufferUsageFlagBits => a

VK_COMMAND_BUFFER_USAGE_RENDER_PASS_CONTINUE_BIT specifies that a secondary command buffer is considered to be entirely inside a render pass. If this is a primary command buffer, then this bit is ignored.

pattern COMMAND_BUFFER_USAGE_SIMULTANEOUS_USE_BIT :: a ~ CommandBufferUsageFlagBits => a

VK_COMMAND_BUFFER_USAGE_SIMULTANEOUS_USE_BIT specifies that a command buffer can be resubmitted to a queue while it is in the pending state, and recorded into multiple primary command buffers.

type CommandBufferUsageFlags = CommandBufferUsageFlagBits

VkCommandBufferUsageFlags - Bitmask of VkCommandBufferUsageFlagBits

Description

VkCommandBufferUsageFlags is a bitmask type for setting a mask of zero or more VkCommandBufferUsageFlagBits.

See Also

VkCommandBufferBeginInfo, VkCommandBufferUsageFlagBits

type QueryControlFlagBits = VkQueryControlFlagBits

VkQueryControlFlagBits - Bitmask specifying constraints on a query

See Also

VkQueryControlFlags

pattern QUERY_CONTROL_PRECISE_BIT :: a ~ QueryControlFlagBits => a

VK_QUERY_CONTROL_PRECISE_BIT specifies the precision of occlusion queries.

type QueryControlFlags = QueryControlFlagBits

VkQueryControlFlags - Bitmask of VkQueryControlFlagBits

Description

VkQueryControlFlags is a bitmask type for setting a mask of zero or more VkQueryControlFlagBits.

See Also

VkCommandBufferInheritanceInfo, VkQueryControlFlagBits, vkCmdBeginQuery, vkCmdBeginQueryIndexedEXT

allocateCommandBuffers :: Device -> CommandBufferAllocateInfo -> IO (Vector CommandBuffer)

vkAllocateCommandBuffers - Allocate command buffers from an existing command pool

Parameters

device is the logical device that owns the command pool.

pAllocateInfo is a pointer to an instance of the VkCommandBufferAllocateInfo structure describing parameters of the allocation.

pCommandBuffers is a pointer to an array of VkCommandBuffer handles in which the resulting command buffer objects are returned. The array must be at least the length specified by the commandBufferCount member of pAllocateInfo. Each allocated command buffer begins in the initial state.

Description

When command buffers are first allocated, they are in the initial state.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pAllocateInfo must be a valid pointer to a valid VkCommandBufferAllocateInfo structure

pCommandBuffers must be a valid pointer to an array of pAllocateInfo::commandBufferCount VkCommandBuffer handles

Host Synchronization

Host access to pAllocateInfo::commandPool must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkCommandBuffer, VkCommandBufferAllocateInfo, VkDevice

beginCommandBuffer :: CommandBuffer -> CommandBufferBeginInfo -> IO ()

vkBeginCommandBuffer - Start recording a command buffer

Parameters

commandBuffer is the handle of the command buffer which is to be put in the recording state.

pBeginInfo is an instance of the VkCommandBufferBeginInfo structure, which defines additional information about how the command buffer begins recording.

Valid Usage

commandBuffer must not be in the recording or pending state.

If commandBuffer was allocated from a VkCommandPool which did not have the VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT flag set, commandBuffer must be in the initial state.

If commandBuffer is a secondary command buffer, the pInheritanceInfo member of pBeginInfo must be a valid VkCommandBufferInheritanceInfo structure

If commandBuffer is a secondary command buffer and either the occlusionQueryEnable member of the pInheritanceInfo member of pBeginInfo is VK_FALSE, or the precise occlusion queries feature is not enabled, the queryFlags member of the pInheritanceInfo member pBeginInfo must not contain VK_QUERY_CONTROL_PRECISE_BIT

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pBeginInfo must be a valid pointer to a valid VkCommandBufferBeginInfo structure

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkCommandBuffer, VkCommandBufferBeginInfo

endCommandBuffer :: CommandBuffer -> IO ()

vkEndCommandBuffer - Finish recording a command buffer

Parameters

commandBuffer is the command buffer to complete recording.

Description

If there was an error during recording, the application will be notified by an unsuccessful return code returned by vkEndCommandBuffer. If the application wishes to further use the command buffer, the command buffer must be reset. The command buffer must have been in the recording state, and is moved to the executable state.

Valid Usage

commandBuffer must be in the recording state.

If commandBuffer is a primary command buffer, there must not be an active render pass instance

All queries made active during the recording of commandBuffer must have been made inactive

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkCommandBuffer

freeCommandBuffers :: Device -> CommandPool -> Vector CommandBuffer -> IO ()

vkFreeCommandBuffers - Free command buffers

Parameters

device is the logical device that owns the command pool.

commandPool is the command pool from which the command buffers were allocated.

commandBufferCount is the length of the pCommandBuffers array.

pCommandBuffers is an array of handles of command buffers to free.

Description

Any primary command buffer that is in the recording or executable state and has any element of pCommandBuffers recorded into it, becomes invalid.

Valid Usage

All elements of pCommandBuffers must not be in the pending state

pCommandBuffers must be a valid pointer to an array of commandBufferCount VkCommandBuffer handles, each element of which must either be a valid handle or NULL

Valid Usage (Implicit)

device must be a valid VkDevice handle

commandPool must be a valid VkCommandPool handle

commandBufferCount must be greater than 0

commandPool must have been created, allocated, or retrieved from device

Each element of pCommandBuffers that is a valid handle must have been created, allocated, or retrieved from commandPool

Host Synchronization

Host access to commandPool must be externally synchronized

Host access to each member of pCommandBuffers must be externally synchronized

See Also

VkCommandBuffer, VkCommandPool, VkDevice

resetCommandBuffer :: CommandBuffer -> CommandBufferResetFlags -> IO ()

vkResetCommandBuffer - Reset a command buffer to the initial state

Parameters

commandBuffer is the command buffer to reset. The command buffer can be in any state other than pending, and is moved into the initial state.

flags is a bitmask of VkCommandBufferResetFlagBits controlling the reset operation.

Description

Any primary command buffer that is in the recording or executable state and has commandBuffer recorded into it, becomes invalid.

Valid Usage

commandBuffer must not be in the pending state

commandBuffer must have been allocated from a pool that was created with the VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

flags must be a valid combination of VkCommandBufferResetFlagBits values

Host Synchronization

Host access to commandBuffer must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkCommandBuffer, VkCommandBufferResetFlags

useCommandBuffer :: CommandBuffer -> CommandBufferBeginInfo -> IO a -> IO a

A safe wrapper for beginCommandBuffer and endCommandBuffer using bracket_

The allocated value must not be returned from the provided computation

withCommandBuffers :: Device -> CommandBufferAllocateInfo -> (Vector CommandBuffer -> IO a) -> IO a

A safe wrapper for allocateCommandBuffers and freeCommandBuffers using bracket

The allocated value must not be returned from the provided computation