Graphics.Vulkan.Core10.Device

Documentation

type DeviceCreateFlags = VkDeviceCreateFlags

VkDeviceCreateFlags - Reserved for future use

Description

VkDeviceCreateFlags is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkDeviceCreateInfo

withCStructDeviceCreateInfo :: DeviceCreateInfo -> (VkDeviceCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkDeviceCreateInfo and marshal a DeviceCreateInfo into it. The VkDeviceCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructDeviceCreateInfo :: VkDeviceCreateInfo -> IO DeviceCreateInfo

A function to read a VkDeviceCreateInfo and all additional structures in the pointer chain into a DeviceCreateInfo.

data DeviceCreateInfo

VkDeviceCreateInfo - Structure specifying parameters of a newly created device

Valid Usage

The queueFamilyIndex member of each element of pQueueCreateInfos must be unique within pQueueCreateInfos

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DEVICE_CREATE_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkDeviceGroupDeviceCreateInfo, VkDeviceMemoryOverallocationCreateInfoAMD, VkPhysicalDevice16BitStorageFeatures, VkPhysicalDevice8BitStorageFeaturesKHR, VkPhysicalDeviceASTCDecodeFeaturesEXT, VkPhysicalDeviceBlendOperationAdvancedFeaturesEXT, VkPhysicalDeviceBufferDeviceAddressFeaturesEXT, VkPhysicalDeviceComputeShaderDerivativesFeaturesNV, VkPhysicalDeviceConditionalRenderingFeaturesEXT, VkPhysicalDeviceCooperativeMatrixFeaturesNV, VkPhysicalDeviceCornerSampledImageFeaturesNV, VkPhysicalDeviceDedicatedAllocationImageAliasingFeaturesNV, VkPhysicalDeviceDepthClipEnableFeaturesEXT, VkPhysicalDeviceDescriptorIndexingFeaturesEXT, VkPhysicalDeviceExclusiveScissorFeaturesNV, VkPhysicalDeviceFeatures2, VkPhysicalDeviceFloat16Int8FeaturesKHR, VkPhysicalDeviceFragmentDensityMapFeaturesEXT, VkPhysicalDeviceFragmentShaderBarycentricFeaturesNV, VkPhysicalDeviceHostQueryResetFeaturesEXT, VkPhysicalDeviceInlineUniformBlockFeaturesEXT, VkPhysicalDeviceMemoryPriorityFeaturesEXT, VkPhysicalDeviceMeshShaderFeaturesNV, VkPhysicalDeviceMultiviewFeatures, VkPhysicalDeviceProtectedMemoryFeatures, VkPhysicalDeviceRepresentativeFragmentTestFeaturesNV, VkPhysicalDeviceSamplerYcbcrConversionFeatures, VkPhysicalDeviceScalarBlockLayoutFeaturesEXT, VkPhysicalDeviceShaderAtomicInt64FeaturesKHR, VkPhysicalDeviceShaderDrawParametersFeatures, VkPhysicalDeviceShaderImageFootprintFeaturesNV, VkPhysicalDeviceShadingRateImageFeaturesNV, VkPhysicalDeviceTransformFeedbackFeaturesEXT, VkPhysicalDeviceVariablePointersFeatures, VkPhysicalDeviceVertexAttributeDivisorFeaturesEXT, VkPhysicalDeviceVulkanMemoryModelFeaturesKHR, or VkPhysicalDeviceYcbcrImageArraysFeaturesEXT

Each sType member in the pNext chain must be unique

flags must be 0

pQueueCreateInfos must be a valid pointer to an array of queueCreateInfoCount valid VkDeviceQueueCreateInfo structures

If enabledLayerCount is not 0, ppEnabledLayerNames must be a valid pointer to an array of enabledLayerCount null-terminated UTF-8 strings

If enabledExtensionCount is not 0, ppEnabledExtensionNames must be a valid pointer to an array of enabledExtensionCount null-terminated UTF-8 strings

If pEnabledFeatures is not NULL, pEnabledFeatures must be a valid pointer to a valid VkPhysicalDeviceFeatures structure

queueCreateInfoCount must be greater than 0

See Also

VkDeviceCreateFlags, VkDeviceQueueCreateInfo, VkPhysicalDeviceFeatures, VkStructureType, vkCreateDevice

Constructors

DeviceCreateInfo

Fields

next :: Maybe SomeVkStruct

flags :: DeviceCreateFlags

queueCreateInfos :: Vector DeviceQueueCreateInfo

enabledLayerNames :: Vector ByteString

enabledExtensionNames :: Vector ByteString

enabledFeatures :: Maybe PhysicalDeviceFeatures

type DeviceQueueCreateFlagBits = VkDeviceQueueCreateFlagBits

VkDeviceQueueCreateFlagBits - Bitmask specifying behavior of the queue

See Also

VkDeviceQueueCreateFlags

pattern DEVICE_QUEUE_CREATE_PROTECTED_BIT :: a ~ DeviceQueueCreateFlagBits => a

VK_DEVICE_QUEUE_CREATE_PROTECTED_BIT specifies that the device queue is a protected-capable queue. If the protected memory feature is not enabled, the VK_DEVICE_QUEUE_CREATE_PROTECTED_BIT bit of flags must not be set.

type DeviceQueueCreateFlags = DeviceQueueCreateFlagBits

VkDeviceQueueCreateFlags - Bitmask of VkDeviceQueueCreateFlagBits

Description

VkDeviceQueueCreateFlags is a bitmask type for setting a mask of zero or more VkDeviceQueueCreateFlagBits.

See Also

VkDeviceQueueCreateFlagBits, VkDeviceQueueCreateInfo, VkDeviceQueueInfo2

withCStructDeviceQueueCreateInfo :: DeviceQueueCreateInfo -> (VkDeviceQueueCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkDeviceQueueCreateInfo and marshal a DeviceQueueCreateInfo into it. The VkDeviceQueueCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructDeviceQueueCreateInfo :: VkDeviceQueueCreateInfo -> IO DeviceQueueCreateInfo

A function to read a VkDeviceQueueCreateInfo and all additional structures in the pointer chain into a DeviceQueueCreateInfo.

data DeviceQueueCreateInfo

VkDeviceQueueCreateInfo - Structure specifying parameters of a newly created device queue

Valid Usage

queueFamilyIndex must be less than pQueueFamilyPropertyCount returned by vkGetPhysicalDeviceQueueFamilyProperties

queueCount must be less than or equal to the queueCount member of the VkQueueFamilyProperties structure, as returned by vkGetPhysicalDeviceQueueFamilyProperties in the pQueueFamilyProperties[queueFamilyIndex]

Each element of pQueuePriorities must be between 0.0 and 1.0 inclusive

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DEVICE_QUEUE_CREATE_INFO

pNext must be NULL or a pointer to a valid instance of VkDeviceQueueGlobalPriorityCreateInfoEXT

flags must be a valid combination of VkDeviceQueueCreateFlagBits values

pQueuePriorities must be a valid pointer to an array of queueCount float values

queueCount must be greater than 0

See Also

VkDeviceCreateInfo, VkDeviceQueueCreateFlags, VkStructureType

Constructors

DeviceQueueCreateInfo

Fields

next :: Maybe SomeVkStruct

flags :: DeviceQueueCreateFlags

queueFamilyIndex :: Word32

queuePriorities :: Vector CFloat

createDevice :: PhysicalDevice -> DeviceCreateInfo -> Maybe AllocationCallbacks -> IO Device

vkCreateDevice - Create a new device instance

Parameters

physicalDevice must be one of the device handles returned from a call to vkEnumeratePhysicalDevices (see Physical Device Enumeration).

pCreateInfo is a pointer to a VkDeviceCreateInfo structure containing information about how to create the device.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pDevice points to a handle in which the created VkDevice is returned.

Description

vkCreateDevice verifies that extensions and features requested in the ppEnabledExtensionNames and pEnabledFeatures members of pCreateInfo, respectively, are supported by the implementation. If any requested extension is not supported, vkCreateDevice must return VK_ERROR_EXTENSION_NOT_PRESENT. If any requested feature is not supported, vkCreateDevice must return VK_ERROR_FEATURE_NOT_PRESENT. Support for extensions can be checked before creating a device by querying vkEnumerateDeviceExtensionProperties. Support for features can similarly be checked by querying vkGetPhysicalDeviceFeatures.

After verifying and enabling the extensions the VkDevice object is created and returned to the application. If a requested extension is only supported by a layer, both the layer and the extension need to be specified at vkCreateInstance time for the creation to succeed.

Multiple logical devices can be created from the same physical device. Logical device creation may fail due to lack of device-specific resources (in addition to the other errors). If that occurs, vkCreateDevice will return VK_ERROR_TOO_MANY_OBJECTS.

Valid Usage

All required extensions for each extension in the VkDeviceCreateInfo::ppEnabledExtensionNames list must also be present in that list.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pCreateInfo must be a valid pointer to a valid VkDeviceCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pDevice must be a valid pointer to a VkDevice handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_INITIALIZATION_FAILED

VK_ERROR_EXTENSION_NOT_PRESENT

VK_ERROR_FEATURE_NOT_PRESENT

VK_ERROR_TOO_MANY_OBJECTS

VK_ERROR_DEVICE_LOST

See Also

VkAllocationCallbacks, VkDevice, VkDeviceCreateInfo, VkPhysicalDevice

destroyDevice :: Device -> Maybe AllocationCallbacks -> IO ()

vkDestroyDevice - Destroy a logical device

Parameters

device is the logical device to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Description

To ensure that no work is active on the device, vkDeviceWaitIdle can be used to gate the destruction of the device. Prior to destroying a device, an application is responsible for destroying/freeing any Vulkan objects that were created using that device as the first parameter of the corresponding vkCreate* or vkAllocate* command.

Note

The lifetime of each of these objects is bound by the lifetime of the VkDevice object. Therefore, to avoid resource leaks, it is critical that an application explicitly free all of these resources prior to calling vkDestroyDevice.

Valid Usage

All child objects created on device must have been destroyed prior to destroying device

If VkAllocationCallbacks were provided when device was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when device was created, pAllocator must be NULL

Valid Usage (Implicit)

If device is not NULL, device must be a valid VkDevice handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

Host Synchronization

Host access to device must be externally synchronized

See Also

VkAllocationCallbacks, VkDevice

withDevice :: PhysicalDevice -> DeviceCreateInfo -> Maybe AllocationCallbacks -> (Device -> IO a) -> IO a

A safe wrapper for createDevice and destroyDevice using bracket

The allocated value must not be returned from the provided computation