Graphics.Vulkan.Core10.Enums.CompareOp

Documentation

newtype CompareOp

VkCompareOp - Stencil comparison function

See Also

PipelineDepthStencilStateCreateInfo, SamplerCreateInfo, StencilOpState

Constructors

CompareOp Int32

Bundled Patterns

pattern COMPARE_OP_NEVER :: CompareOp
COMPARE_OP_NEVER specifies that the test never passes.

pattern COMPARE_OP_LESS :: CompareOp
COMPARE_OP_LESS specifies that the test passes when R < S.

pattern COMPARE_OP_EQUAL :: CompareOp
COMPARE_OP_EQUAL specifies that the test passes when R = S.

pattern COMPARE_OP_LESS_OR_EQUAL :: CompareOp
COMPARE_OP_LESS_OR_EQUAL specifies that the test passes when R ≤ S.

pattern COMPARE_OP_GREATER :: CompareOp
COMPARE_OP_GREATER specifies that the test passes when R > S.

pattern COMPARE_OP_NOT_EQUAL :: CompareOp
COMPARE_OP_NOT_EQUAL specifies that the test passes when R ≠ S.

pattern COMPARE_OP_GREATER_OR_EQUAL :: CompareOp
COMPARE_OP_GREATER_OR_EQUAL specifies that the test passes when R ≥ S.

pattern COMPARE_OP_ALWAYS :: CompareOp
COMPARE_OP_ALWAYS specifies that the test always passes.