Graphics.Vulkan.Core10.Enums.ComponentSwizzle

Documentation

newtype ComponentSwizzle

VkComponentSwizzle - Specify how a component is swizzled

Description

Setting the identity swizzle on a component is equivalent to setting the identity mapping on that component. That is:

Component 	 Identity Mapping

components.r 	 COMPONENT_SWIZZLE_R

components.g 	 COMPONENT_SWIZZLE_G

components.b 	 COMPONENT_SWIZZLE_B

components.a 	 COMPONENT_SWIZZLE_A

Component Mappings Equivalent To COMPONENT_SWIZZLE_IDENTITY

See Also

ComponentMapping

Constructors

ComponentSwizzle Int32

Bundled Patterns

pattern COMPONENT_SWIZZLE_IDENTITY :: ComponentSwizzle
COMPONENT_SWIZZLE_IDENTITY specifies that the component is set to the identity swizzle.

pattern COMPONENT_SWIZZLE_ZERO :: ComponentSwizzle
COMPONENT_SWIZZLE_ZERO specifies that the component is set to zero.

pattern COMPONENT_SWIZZLE_ONE :: ComponentSwizzle
COMPONENT_SWIZZLE_ONE specifies that the component is set to either 1 or 1.0, depending on whether the type of the image view format is integer or floating-point respectively, as determined by the Format Definition section for each Format.

pattern COMPONENT_SWIZZLE_R :: ComponentSwizzle
COMPONENT_SWIZZLE_R specifies that the component is set to the value of the R component of the image.

pattern COMPONENT_SWIZZLE_G :: ComponentSwizzle
COMPONENT_SWIZZLE_G specifies that the component is set to the value of the G component of the image.

pattern COMPONENT_SWIZZLE_B :: ComponentSwizzle
COMPONENT_SWIZZLE_B specifies that the component is set to the value of the B component of the image.

pattern COMPONENT_SWIZZLE_A :: ComponentSwizzle
COMPONENT_SWIZZLE_A specifies that the component is set to the value of the A component of the image.