Graphics.Vulkan.Core10.Enums.QueryType

Documentation

newtype QueryType

VkQueryType - Specify the type of queries managed by a query pool

See Also

QueryPoolCreateInfo, cmdWriteAccelerationStructuresPropertiesNV

Constructors

QueryType Int32

Bundled Patterns

pattern QUERY_TYPE_OCCLUSION :: QueryType
QUERY_TYPE_OCCLUSION specifies an occlusion query.

pattern QUERY_TYPE_PIPELINE_STATISTICS :: QueryType
QUERY_TYPE_PIPELINE_STATISTICS specifies a pipeline statistics query.

pattern QUERY_TYPE_TIMESTAMP :: QueryType
QUERY_TYPE_TIMESTAMP specifies a timestamp query.

pattern QUERY_TYPE_PERFORMANCE_QUERY_INTEL :: QueryType
QUERY_TYPE_PERFORMANCE_QUERY_INTEL specifies a Intel performance query.

pattern QUERY_TYPE_ACCELERATION_STRUCTURE_COMPACTED_SIZE_NV :: QueryType
QUERY_TYPE_ACCELERATION_STRUCTURE_COMPACTED_SIZE_NV specifies a ray tracing acceleration structure size query.

pattern QUERY_TYPE_PERFORMANCE_QUERY_KHR :: QueryType
QUERY_TYPE_PERFORMANCE_QUERY_KHR specifies a performance query.

pattern QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT :: QueryType
QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT specifies a transform feedback query.