Graphics.Vulkan.Core10.Enums.SampleCountFlagBits

Documentation

newtype SampleCountFlagBits

VkSampleCountFlagBits - Bitmask specifying sample counts supported for an image used for storage operations

See Also

AttachmentDescription, AttachmentDescription2, FramebufferMixedSamplesCombinationNV, ImageCreateInfo, PhysicalDeviceSparseImageFormatInfo2, PipelineMultisampleStateCreateInfo, SampleCountFlags, SampleLocationsInfoEXT, getPhysicalDeviceMultisamplePropertiesEXT, getPhysicalDeviceSparseImageFormatProperties

Constructors

SampleCountFlagBits Flags

Bundled Patterns

pattern SAMPLE_COUNT_1_BIT :: SampleCountFlagBits
SAMPLE_COUNT_1_BIT specifies an image with one sample per pixel.

pattern SAMPLE_COUNT_2_BIT :: SampleCountFlagBits
SAMPLE_COUNT_2_BIT specifies an image with 2 samples per pixel.

pattern SAMPLE_COUNT_4_BIT :: SampleCountFlagBits
SAMPLE_COUNT_4_BIT specifies an image with 4 samples per pixel.

pattern SAMPLE_COUNT_8_BIT :: SampleCountFlagBits
SAMPLE_COUNT_8_BIT specifies an image with 8 samples per pixel.

pattern SAMPLE_COUNT_16_BIT :: SampleCountFlagBits
SAMPLE_COUNT_16_BIT specifies an image with 16 samples per pixel.

pattern SAMPLE_COUNT_32_BIT :: SampleCountFlagBits
SAMPLE_COUNT_32_BIT specifies an image with 32 samples per pixel.

pattern SAMPLE_COUNT_64_BIT :: SampleCountFlagBits
SAMPLE_COUNT_64_BIT specifies an image with 64 samples per pixel.

type SampleCountFlags = SampleCountFlagBits