Graphics.Vulkan.Core10.ImageView

Documentation

withCStructComponentMapping :: ComponentMapping -> (VkComponentMapping -> IO a) -> IO a

A function to temporarily allocate memory for a VkComponentMapping and marshal a ComponentMapping into it. The VkComponentMapping is only valid inside the provided computation and must not be returned out of it.

fromCStructComponentMapping :: VkComponentMapping -> IO ComponentMapping

A function to read a VkComponentMapping and all additional structures in the pointer chain into a ComponentMapping.

data ComponentMapping

VkComponentMapping - Structure specifying a color component mapping

Valid Usage (Implicit)

See Also

VkAndroidHardwareBufferFormatPropertiesANDROID, VkComponentSwizzle, VkImageViewCreateInfo, VkSamplerYcbcrConversionCreateInfo

Constructors

ComponentMapping

Fields

r :: ComponentSwizzle

g :: ComponentSwizzle

b :: ComponentSwizzle

a :: ComponentSwizzle

type ComponentSwizzle = VkComponentSwizzle

VkComponentSwizzle - Specify how a component is swizzled

Description

Setting the identity swizzle on a component is equivalent to setting the identity mapping on that component. That is:

+-----------------------------------+-----------------------------------+
| Component                         | Identity Mapping                  |
+===================================+===================================+
| @components.r@                    | 'Graphics.Vulkan.C.Core10.ImageVi |
|                                   | ew.VK_COMPONENT_SWIZZLE_R'        |
+-----------------------------------+-----------------------------------+
| @components.g@                    | 'Graphics.Vulkan.C.Core10.ImageVi |
|                                   | ew.VK_COMPONENT_SWIZZLE_G'        |
+-----------------------------------+-----------------------------------+
| @components.b@                    | 'Graphics.Vulkan.C.Core10.ImageVi |
|                                   | ew.VK_COMPONENT_SWIZZLE_B'        |
+-----------------------------------+-----------------------------------+
| @components.a@                    | 'Graphics.Vulkan.C.Core10.ImageVi |
|                                   | ew.VK_COMPONENT_SWIZZLE_A'        |
+-----------------------------------+-----------------------------------+

Component Mappings Equivalent To
'Graphics.Vulkan.C.Core10.ImageView.VK_COMPONENT_SWIZZLE_IDENTITY'

See Also

VkComponentMapping

pattern COMPONENT_SWIZZLE_IDENTITY :: a ~ ComponentSwizzle => a

VK_COMPONENT_SWIZZLE_IDENTITY specifies that the component is set to the identity swizzle.

pattern COMPONENT_SWIZZLE_ZERO :: a ~ ComponentSwizzle => a

VK_COMPONENT_SWIZZLE_ZERO specifies that the component is set to zero.

pattern COMPONENT_SWIZZLE_ONE :: a ~ ComponentSwizzle => a

VK_COMPONENT_SWIZZLE_ONE specifies that the component is set to either 1 or 1.0, depending on whether the type of the image view format is integer or floating-point respectively, as determined by the Format Definition section for each VkFormat.

pattern COMPONENT_SWIZZLE_R :: a ~ ComponentSwizzle => a

VK_COMPONENT_SWIZZLE_R specifies that the component is set to the value of the R component of the image.

pattern COMPONENT_SWIZZLE_G :: a ~ ComponentSwizzle => a

VK_COMPONENT_SWIZZLE_G specifies that the component is set to the value of the G component of the image.

pattern COMPONENT_SWIZZLE_B :: a ~ ComponentSwizzle => a

VK_COMPONENT_SWIZZLE_B specifies that the component is set to the value of the B component of the image.

pattern COMPONENT_SWIZZLE_A :: a ~ ComponentSwizzle => a

VK_COMPONENT_SWIZZLE_A specifies that the component is set to the value of the A component of the image.

withCStructImageSubresourceRange :: ImageSubresourceRange -> (VkImageSubresourceRange -> IO a) -> IO a

A function to temporarily allocate memory for a VkImageSubresourceRange and marshal a ImageSubresourceRange into it. The VkImageSubresourceRange is only valid inside the provided computation and must not be returned out of it.

fromCStructImageSubresourceRange :: VkImageSubresourceRange -> IO ImageSubresourceRange

A function to read a VkImageSubresourceRange and all additional structures in the pointer chain into a ImageSubresourceRange.

data ImageSubresourceRange

VkImageSubresourceRange - Structure specifying an image subresource range

Description

The number of mipmap levels and array layers must be a subset of the image subresources in the image. If an application wants to use all mip levels or layers in an image after the baseMipLevel or baseArrayLayer, it can set levelCount and layerCount to the special values VK_REMAINING_MIP_LEVELS and VK_REMAINING_ARRAY_LAYERS without knowing the exact number of mip levels or layers.

For cube and cube array image views, the layers of the image view starting at baseArrayLayer correspond to faces in the order +X, -X, +Y, -Y, +Z, -Z. For cube arrays, each set of six sequential layers is a single cube, so the number of cube maps in a cube map array view is layerCount / 6, and image array layer (baseArrayLayer + i) is face index (i mod 6) of cube i / 6. If the number of layers in the view, whether set explicitly in layerCount or implied by VK_REMAINING_ARRAY_LAYERS, is not a multiple of 6, the last cube map in the array must not be accessed.

aspectMask must be only VK_IMAGE_ASPECT_COLOR_BIT, VK_IMAGE_ASPECT_DEPTH_BIT or VK_IMAGE_ASPECT_STENCIL_BIT if format is a color, depth-only or stencil-only format, respectively. If using a depth/stencil format with both depth and stencil components, aspectMask must include at least one of VK_IMAGE_ASPECT_DEPTH_BIT and VK_IMAGE_ASPECT_STENCIL_BIT, and can include both.

When using an image view of a depth/stencil image to populate a descriptor set (e.g. for sampling in the shader, or for use as an input attachment), the aspectMask must only include one bit and selects whether the image view is used for depth reads (i.e. using a floating-point sampler or input attachment in the shader) or stencil reads (i.e. using an unsigned integer sampler or input attachment in the shader). When an image view of a depth/stencil image is used as a depth/stencil framebuffer attachment, the aspectMask is ignored and both depth and stencil image subresources are used.

The components member is of type VkComponentMapping, and describes a remapping from components of the image to components of the vector returned by shader image instructions. This remapping must be identity for storage image descriptors, input attachment descriptors, and framebuffer attachments.

Valid Usage

If levelCount is not VK_REMAINING_MIP_LEVELS, it must be greater than 0

If layerCount is not VK_REMAINING_ARRAY_LAYERS, it must be greater than 0

Valid Usage (Implicit)

aspectMask must be a valid combination of VkImageAspectFlagBits values

aspectMask must not be 0

See Also

VkImageAspectFlags, VkImageMemoryBarrier, VkImageViewCreateInfo, vkCmdClearColorImage, vkCmdClearDepthStencilImage

Constructors

ImageSubresourceRange

Fields

aspectMask :: ImageAspectFlags

baseMipLevel :: Word32

levelCount :: Word32

baseArrayLayer :: Word32

layerCount :: Word32

type ImageView = VkImageView

VkImageView - Opaque handle to an image view object

See Also

VkDescriptorImageInfo, VkFramebufferCreateInfo, VkImageViewHandleInfoNVX, vkCmdBindShadingRateImageNV, vkCreateImageView, vkDestroyImageView

type ImageViewCreateFlagBits = VkImageViewCreateFlagBits

VkImageViewCreateFlagBits - Bitmask specifying additional parameters of an image view

See Also

VkImageViewCreateFlags

pattern IMAGE_VIEW_CREATE_FRAGMENT_DENSITY_MAP_DYNAMIC_BIT_EXT :: a ~ ImageViewCreateFlagBits => a

type ImageViewCreateFlags = ImageViewCreateFlagBits

VkImageViewCreateFlags - Reserved for future use

Description

VkImageViewCreateFlags is a bitmask type for setting a mask of zero or more VkImageViewCreateFlagBits.

See Also

VkImageViewCreateFlagBits, VkImageViewCreateInfo

withCStructImageViewCreateInfo :: ImageViewCreateInfo -> (VkImageViewCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkImageViewCreateInfo and marshal a ImageViewCreateInfo into it. The VkImageViewCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructImageViewCreateInfo :: VkImageViewCreateInfo -> IO ImageViewCreateInfo

A function to read a VkImageViewCreateInfo and all additional structures in the pointer chain into a ImageViewCreateInfo.

data ImageViewCreateInfo

VkImageViewCreateInfo - Structure specifying parameters of a newly created image view

Description

Some of the image creation parameters are inherited by the view. In particular, image view creation inherits the implicit parameter usage specifying the allowed usages of the image view that, by default, takes the value of the corresponding usage parameter specified in VkImageCreateInfo at image creation time.

If image was created with the VK_IMAGE_CREATE_MUTABLE_FORMAT_BIT flag, format can be different from the image’s format, but if they are not equal they must be compatible. Image format compatibility is defined in the Format Compatibility Classes section. Views of compatible formats will have the same mapping between texel coordinates and memory locations irrespective of the format, with only the interpretation of the bit pattern changing.

Note

Values intended to be used with one view format may not be exactly preserved when written or read through a different format. For example, an integer value that happens to have the bit pattern of a floating point denorm or NaN may be flushed or canonicalized when written or read through a view with a floating point format. Similarly, a value written through a signed normalized format that has a bit pattern exactly equal to -2b may be changed to -2b + 1 as described in Conversion from Normalized Fixed-Point to Floating-Point.

+---------+------------------------+-----------------------------------+
| Dim,    | Image parameters       | View parameters                   |
| Arrayed |                        |                                   |
| ,       |                        |                                   |
| MS      |                        |                                   |
+=========+========================+===================================+
|         | @imageType@ =          | @baseArrayLayer@ and @layerCount@ |
|         | ci.@imageType@         | are members of the                |
|         | @width@ =              | @subresourceRange@ member.        |
|         | ci.@extent.width@      |                                   |
|         | @height@ =             |                                   |
|         | ci.@extent.height@     |                                   |
|         | @depth@ =              |                                   |
|         | ci.@extent.depth@      |                                   |
|         | @arrayLayers@ =        |                                   |
|         | ci.@arrayLayers@       |                                   |
|         | @samples@ =            |                                   |
|         | ci.@samples@           |                                   |
|         | @flags@ = ci.@flags@   |                                   |
|         | where ci is the        |                                   |
|         | 'Graphics.Vulkan.C.Cor |                                   |
|         | e10.Image.VkImageCreat |                                   |
|         | eInfo'                 |                                   |
|         | used to create         |                                   |
|         | @image@.               |                                   |
+---------+------------------------+-----------------------------------+
| __1D,   | @imageType@ =          | @viewType@ =                      |
| 0, 0__  | 'Graphics.Vulkan.C.Cor | 'Graphics.Vulkan.C.Core10.ImageVi |
|         | e10.DeviceInitializati | ew.VK_IMAGE_VIEW_TYPE_1D'         |
|         | on.VK_IMAGE_TYPE_1D'   | @baseArrayLayer@ ≥ 0              |
|         | @width@ ≥ 1            | @layerCount@ = 1                  |
|         | @height@ = 1           |                                   |
|         | @depth@ = 1            |                                   |
|         | @arrayLayers@ ≥ 1      |                                   |
|         | @samples@ = 1          |                                   |
+---------+------------------------+-----------------------------------+
| __1D,   | @imageType@ =          | @viewType@ =                      |
| 1, 0__  | 'Graphics.Vulkan.C.Cor | 'Graphics.Vulkan.C.Core10.ImageVi |
|         | e10.DeviceInitializati | ew.VK_IMAGE_VIEW_TYPE_1D_ARRAY'   |
|         | on.VK_IMAGE_TYPE_1D'   | @baseArrayLayer@ ≥ 0              |
|         | @width@ ≥ 1            | @layerCount@ ≥ 1                  |
|         | @height@ = 1           |                                   |
|         | @depth@ = 1            |                                   |
|         | @arrayLayers@ ≥ 1      |                                   |
|         | @samples@ = 1          |                                   |
+---------+------------------------+-----------------------------------+
| __2D,   | @imageType@ =          | @viewType@ =                      |
| 0, 0__  | 'Graphics.Vulkan.C.Cor | 'Graphics.Vulkan.C.Core10.ImageVi |
|         | e10.DeviceInitializati | ew.VK_IMAGE_VIEW_TYPE_2D'         |
|         | on.VK_IMAGE_TYPE_2D'   | @baseArrayLayer@ ≥ 0              |
|         | @width@ ≥ 1            | @layerCount@ = 1                  |
|         | @height@ ≥ 1           |                                   |
|         | @depth@ = 1            |                                   |
|         | @arrayLayers@ ≥ 1      |                                   |
|         | @samples@ = 1          |                                   |
+---------+------------------------+-----------------------------------+
| __2D,   | @imageType@ =          | @viewType@ =                      |
| 1, 0__  | 'Graphics.Vulkan.C.Cor | 'Graphics.Vulkan.C.Core10.ImageVi |
|         | e10.DeviceInitializati | ew.VK_IMAGE_VIEW_TYPE_2D_ARRAY'   |
|         | on.VK_IMAGE_TYPE_2D'   | @baseArrayLayer@ ≥ 0              |
|         | @width@ ≥ 1            | @layerCount@ ≥ 1                  |
|         | @height@ ≥ 1           |                                   |
|         | @depth@ = 1            |                                   |
|         | @arrayLayers@ ≥ 1      |                                   |
|         | @samples@ = 1          |                                   |
+---------+------------------------+-----------------------------------+
| __2D,   | @imageType@ =          | @viewType@ =                      |
| 0, 1__  | 'Graphics.Vulkan.C.Cor | 'Graphics.Vulkan.C.Core10.ImageVi |
|         | e10.DeviceInitializati | ew.VK_IMAGE_VIEW_TYPE_2D'         |
|         | on.VK_IMAGE_TYPE_2D'   | @baseArrayLayer@ ≥ 0              |
|         | @width@ ≥ 1            | @layerCount@ = 1                  |
|         | @height@ ≥ 1           |                                   |
|         | @depth@ = 1            |                                   |
|         | @arrayLayers@ ≥ 1      |                                   |
|         | @samples@ > 1          |                                   |
+---------+------------------------+-----------------------------------+
| __2D,   | @imageType@ =          | @viewType@ =                      |
| 1, 1__  | 'Graphics.Vulkan.C.Cor | 'Graphics.Vulkan.C.Core10.ImageVi |
|         | e10.DeviceInitializati | ew.VK_IMAGE_VIEW_TYPE_2D_ARRAY'   |
|         | on.VK_IMAGE_TYPE_2D'   | @baseArrayLayer@ ≥ 0              |
|         | @width@ ≥ 1            | @layerCount@ ≥ 1                  |
|         | @height@ ≥ 1           |                                   |
|         | @depth@ = 1            |                                   |
|         | @arrayLayers@ ≥ 1      |                                   |
|         | @samples@ > 1          |                                   |
+---------+------------------------+-----------------------------------+
| __CUBE, | @imageType@ =          | @viewType@ =                      |
| 0, 0__  | 'Graphics.Vulkan.C.Cor | 'Graphics.Vulkan.C.Core10.ImageVi |
|         | e10.DeviceInitializati | ew.VK_IMAGE_VIEW_TYPE_CUBE'       |
|         | on.VK_IMAGE_TYPE_2D'   | @baseArrayLayer@ ≥ 0              |
|         | @width@ ≥ 1            | @layerCount@ = 6                  |
|         | @height@ = @width@     |                                   |
|         | @depth@ = 1            |                                   |
|         | @arrayLayers@ ≥ 6      |                                   |
|         | @samples@ = 1          |                                   |
|         | @flags@ includes       |                                   |
|         | 'Graphics.Vulkan.C.Cor |                                   |
|         | e10.DeviceInitializati |                                   |
|         | on.VK_IMAGE_CREATE_CUB |                                   |
|         | E_COMPATIBLE_BIT'      |                                   |
+---------+------------------------+-----------------------------------+
| __CUBE, | @imageType@ =          | @viewType@ =                      |
| 1, 0__  | 'Graphics.Vulkan.C.Cor | 'Graphics.Vulkan.C.Core10.ImageVi |
|         | e10.DeviceInitializati | ew.VK_IMAGE_VIEW_TYPE_CUBE_ARRAY' |
|         | on.VK_IMAGE_TYPE_2D'   | @baseArrayLayer@ ≥ 0              |
|         | @width@ ≥ 1            | @layerCount@ = 6 × /N/, /N/ ≥ 1   |
|         | @height@ = width       |                                   |
|         | @depth@ = 1            |                                   |
|         | /N/ ≥ 1                |                                   |
|         | @arrayLayers@ ≥ 6 ×    |                                   |
|         | /N/                    |                                   |
|         | @samples@ = 1          |                                   |
|         | @flags@ includes       |                                   |
|         | 'Graphics.Vulkan.C.Cor |                                   |
|         | e10.DeviceInitializati |                                   |
|         | on.VK_IMAGE_CREATE_CUB |                                   |
|         | E_COMPATIBLE_BIT'      |                                   |
+---------+------------------------+-----------------------------------+
| __3D,   | @imageType@ =          | @viewType@ =                      |
| 0, 0__  | 'Graphics.Vulkan.C.Cor | 'Graphics.Vulkan.C.Core10.ImageVi |
|         | e10.DeviceInitializati | ew.VK_IMAGE_VIEW_TYPE_3D'         |
|         | on.VK_IMAGE_TYPE_3D'   | @baseArrayLayer@ = 0              |
|         | @width@ ≥ 1            | @layerCount@ = 1                  |
|         | @height@ ≥ 1           |                                   |
|         | @depth@ ≥ 1            |                                   |
|         | @arrayLayers@ = 1      |                                   |
|         | @samples@ = 1          |                                   |
+---------+------------------------+-----------------------------------+

Image and image view parameter compatibility requirements

Valid Usage

If image was not created with VK_IMAGE_CREATE_CUBE_COMPATIBLE_BIT then viewType must not be VK_IMAGE_VIEW_TYPE_CUBE or VK_IMAGE_VIEW_TYPE_CUBE_ARRAY

If the image cubemap arrays feature is not enabled, viewType must not be VK_IMAGE_VIEW_TYPE_CUBE_ARRAY

image must have been created with a usage value containing at least one of VK_IMAGE_USAGE_SAMPLED_BIT, VK_IMAGE_USAGE_STORAGE_BIT, VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT, VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT, or VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT

The format features of the resultant image view must contain at least one bit.

If usage contains VK_IMAGE_USAGE_SAMPLED_BIT, then the format features of the resultant image view must contain VK_FORMAT_FEATURE_SAMPLED_IMAGE_BIT.

If usage contains VK_IMAGE_USAGE_STORAGE_BIT, then the image view’s format features must contain VK_FORMAT_FEATURE_STORAGE_IMAGE_BIT.

If usage contains VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT, then the image view’s format features must contain VK_FORMAT_FEATURE_COLOR_ATTACHMENT_BIT.

If usage contains VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT, then the image view’s format features must contain VK_FORMAT_FEATURE_DEPTH_STENCIL_ATTACHMENT_BIT.

If usage contains VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT, then the image view’s format features must contain at least one of VK_FORMAT_FEATURE_COLOR_ATTACHMENT_BIT or VK_FORMAT_FEATURE_DEPTH_STENCIL_ATTACHMENT_BIT.

subresourceRange.baseMipLevel must be less than the mipLevels specified in VkImageCreateInfo when image was created

If subresourceRange.levelCount is not VK_REMAINING_MIP_LEVELS, subresourceRange.baseMipLevel + subresourceRange.levelCount must be less than or equal to the mipLevels specified in VkImageCreateInfo when image was created

subresourceRange.baseArrayLayer must be less than the arrayLayers specified in VkImageCreateInfo when image was created

If subresourceRange.layerCount is not VK_REMAINING_ARRAY_LAYERS, subresourceRange.baseArrayLayer + subresourceRange.layerCount must be less than or equal to the arrayLayers specified in VkImageCreateInfo when image was created

If image was created with the VK_IMAGE_CREATE_MUTABLE_FORMAT_BIT flag, format must be compatible with the format used to create image, as defined in Format Compatibility Classes

If image was not created with the VK_IMAGE_CREATE_MUTABLE_FORMAT_BIT flag, format must be identical to the format used to create image

If image is non-sparse then it must be bound completely and contiguously to a single VkDeviceMemory object

subresourceRange and viewType must be compatible with the image, as described in the compatibility table ifdef::VK_VERSION_1_1,VK_KHR_maintenance2

If the pNext chain includes an instance of VkImageViewUsageCreateInfo, its usage member must not include any bits that were not set in the usage member of the VkImageCreateInfo structure used to create image. endif::VK_VERSION_1_1,VK_KHR_maintenance2

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkImageViewASTCDecodeModeEXT, VkImageViewUsageCreateInfo, or VkSamplerYcbcrConversionInfo

Each sType member in the pNext chain must be unique

flags must be a valid combination of VkImageViewCreateFlagBits values

image must be a valid VkImage handle

viewType must be a valid VkImageViewType value

format must be a valid VkFormat value

components must be a valid VkComponentMapping structure

subresourceRange must be a valid VkImageSubresourceRange structure

See Also

VkComponentMapping, VkFormat, VkImage, VkImageSubresourceRange, VkImageViewCreateFlags, VkImageViewType, VkStructureType, vkCreateImageView

Constructors

ImageViewCreateInfo

Fields

next :: Maybe SomeVkStruct

flags :: ImageViewCreateFlags

image :: Image

viewType :: ImageViewType

format :: Format

components :: ComponentMapping

subresourceRange :: ImageSubresourceRange

type ImageViewType = VkImageViewType

VkImageViewType - Image view types

Description

The exact image view type is partially implicit, based on the image’s type and sample count, as well as the view creation parameters as described in the image view compatibility table for vkCreateImageView. This table also shows which SPIR-V OpTypeImage Dim and Arrayed parameters correspond to each image view type.

See Also

VkImageViewCreateInfo, VkPhysicalDeviceImageViewImageFormatInfoEXT

pattern IMAGE_VIEW_TYPE_1D :: a ~ ImageViewType => a

pattern IMAGE_VIEW_TYPE_2D :: a ~ ImageViewType => a

pattern IMAGE_VIEW_TYPE_3D :: a ~ ImageViewType => a

pattern IMAGE_VIEW_TYPE_CUBE :: a ~ ImageViewType => a

pattern IMAGE_VIEW_TYPE_1D_ARRAY :: a ~ ImageViewType => a

pattern IMAGE_VIEW_TYPE_2D_ARRAY :: a ~ ImageViewType => a

pattern IMAGE_VIEW_TYPE_CUBE_ARRAY :: a ~ ImageViewType => a

createImageView :: Device -> ImageViewCreateInfo -> Maybe AllocationCallbacks -> IO ImageView

vkCreateImageView - Create an image view from an existing image

Parameters

device is the logical device that creates the image view.

pCreateInfo is a pointer to an instance of the VkImageViewCreateInfo structure containing parameters to be used to create the image view.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pView points to a VkImageView handle in which the resulting image view object is returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkImageViewCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pView must be a valid pointer to a VkImageView handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDevice, VkImageView, VkImageViewCreateInfo

destroyImageView :: Device -> ImageView -> Maybe AllocationCallbacks -> IO ()

vkDestroyImageView - Destroy an image view object

Parameters

device is the logical device that destroys the image view.

imageView is the image view to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to imageView must have completed execution

If VkAllocationCallbacks were provided when imageView was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when imageView was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid VkDevice handle

If imageView is not VK_NULL_HANDLE, imageView must be a valid VkImageView handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

If imageView is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to imageView must be externally synchronized

See Also

VkAllocationCallbacks, VkDevice, VkImageView

withImageView :: Device -> ImageViewCreateInfo -> Maybe AllocationCallbacks -> (ImageView -> IO a) -> IO a

A safe wrapper for createImageView and destroyImageView using bracket

The allocated value must not be returned from the provided computation