Graphics.Vulkan.Core10.LayerDiscovery

Documentation

withCStructLayerProperties :: LayerProperties -> (VkLayerProperties -> IO a) -> IO a

A function to temporarily allocate memory for a VkLayerProperties and marshal a LayerProperties into it. The VkLayerProperties is only valid inside the provided computation and must not be returned out of it.

fromCStructLayerProperties :: VkLayerProperties -> IO LayerProperties

A function to read a VkLayerProperties and all additional structures in the pointer chain into a LayerProperties.

data LayerProperties

VkLayerProperties - Structure specifying layer properties

See Also

vkEnumerateDeviceLayerProperties, vkEnumerateInstanceLayerProperties

Constructors

LayerProperties

Fields

layerName :: ByteString

specVersion :: Word32

implementationVersion :: Word32

description :: ByteString

getNumDeviceLayerProperties :: PhysicalDevice -> IO (VkResult, Word32)

vkEnumerateDeviceLayerProperties - Returns properties of available physical device layers

Parameters

pPropertyCount is a pointer to an integer related to the number of layer properties available or queried.

pProperties is either NULL or a pointer to an array of VkLayerProperties structures.

Description

If pProperties is NULL, then the number of layer properties available is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If pPropertyCount is less than the number of layer properties available, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of layers available, VK_INCOMPLETE will be returned instead of VK_SUCCESS, to indicate that not all the available layer properties were returned.

The list of layers enumerated by vkEnumerateDeviceLayerProperties must be exactly the sequence of layers enabled for the instance. The members of VkLayerProperties for each enumerated layer must be the same as the properties when the layer was enumerated by vkEnumerateInstanceLayerProperties.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkLayerProperties structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkLayerProperties, VkPhysicalDevice

enumerateDeviceLayerProperties :: PhysicalDevice -> Word32 -> IO (VkResult, Vector LayerProperties)

vkEnumerateDeviceLayerProperties - Returns properties of available physical device layers

Parameters

pPropertyCount is a pointer to an integer related to the number of layer properties available or queried.

pProperties is either NULL or a pointer to an array of VkLayerProperties structures.

Description

If pProperties is NULL, then the number of layer properties available is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If pPropertyCount is less than the number of layer properties available, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of layers available, VK_INCOMPLETE will be returned instead of VK_SUCCESS, to indicate that not all the available layer properties were returned.

The list of layers enumerated by vkEnumerateDeviceLayerProperties must be exactly the sequence of layers enabled for the instance. The members of VkLayerProperties for each enumerated layer must be the same as the properties when the layer was enumerated by vkEnumerateInstanceLayerProperties.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkLayerProperties structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkLayerProperties, VkPhysicalDevice

enumerateAllDeviceLayerProperties :: PhysicalDevice -> IO (Vector LayerProperties)

Returns all the values available from enumerateDeviceLayerProperties.

getNumInstanceLayerProperties :: IO (VkResult, Word32)

vkEnumerateInstanceLayerProperties - Returns up to requested number of global layer properties

Parameters

pPropertyCount is a pointer to an integer related to the number of layer properties available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkLayerProperties structures.

Description

If pProperties is NULL, then the number of layer properties available is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If pPropertyCount is less than the number of layer properties available, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of layers available, VK_INCOMPLETE will be returned instead of VK_SUCCESS, to indicate that not all the available layer properties were returned.

The list of available layers may change at any time due to actions outside of the Vulkan implementation, so two calls to vkEnumerateInstanceLayerProperties with the same parameters may return different results, or retrieve different pPropertyCount values or pProperties contents. Once an instance has been created, the layers enabled for that instance will continue to be enabled and valid for the lifetime of that instance, even if some of them become unavailable for future instances.

Valid Usage (Implicit)

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkLayerProperties structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkLayerProperties

enumerateInstanceLayerProperties :: Word32 -> IO (VkResult, Vector LayerProperties)

vkEnumerateInstanceLayerProperties - Returns up to requested number of global layer properties

Parameters

pPropertyCount is a pointer to an integer related to the number of layer properties available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkLayerProperties structures.

Description

If pProperties is NULL, then the number of layer properties available is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If pPropertyCount is less than the number of layer properties available, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of layers available, VK_INCOMPLETE will be returned instead of VK_SUCCESS, to indicate that not all the available layer properties were returned.

The list of available layers may change at any time due to actions outside of the Vulkan implementation, so two calls to vkEnumerateInstanceLayerProperties with the same parameters may return different results, or retrieve different pPropertyCount values or pProperties contents. Once an instance has been created, the layers enabled for that instance will continue to be enabled and valid for the lifetime of that instance, even if some of them become unavailable for future instances.

Valid Usage (Implicit)

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkLayerProperties structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkLayerProperties

enumerateAllInstanceLayerProperties :: IO (Vector LayerProperties)

Returns all the values available from enumerateInstanceLayerProperties.