Graphics.Vulkan.Core10.OtherTypes

Documentation

data MemoryBarrier

VkMemoryBarrier - Structure specifying a global memory barrier

Description

The first access scope is limited to access types in the source access mask specified by srcAccessMask.

The second access scope is limited to access types in the destination access mask specified by dstAccessMask.

Valid Usage (Implicit)

See Also

AccessFlags, StructureType, cmdPipelineBarrier, cmdWaitEvents

Constructors

MemoryBarrier

Fields

srcAccessMask :: AccessFlags

srcAccessMask must be a valid combination of AccessFlagBits values

dstAccessMask :: AccessFlags

dstAccessMask must be a valid combination of AccessFlagBits values

data BufferMemoryBarrier

VkBufferMemoryBarrier - Structure specifying a buffer memory barrier

Description

The first access scope is limited to access to memory through the specified buffer range, via access types in the source access mask specified by srcAccessMask. If srcAccessMask includes ACCESS_HOST_WRITE_BIT, memory writes performed by that access type are also made visible, as that access type is not performed through a resource.

The second access scope is limited to access to memory through the specified buffer range, via access types in the destination access mask. specified by dstAccessMask. If dstAccessMask includes ACCESS_HOST_WRITE_BIT or ACCESS_HOST_READ_BIT, available memory writes are also made visible to accesses of those types, as those access types are not performed through a resource.

If srcQueueFamilyIndex is not equal to dstQueueFamilyIndex, and srcQueueFamilyIndex is equal to the current queue family, then the memory barrier defines a queue family release operation for the specified buffer range, and the second access scope includes no access, as if dstAccessMask was 0.

If dstQueueFamilyIndex is not equal to srcQueueFamilyIndex, and dstQueueFamilyIndex is equal to the current queue family, then the memory barrier defines a queue family acquire operation for the specified buffer range, and the first access scope includes no access, as if srcAccessMask was 0.

Valid Usage

offset must be less than the size of Buffer

If size is not equal to WHOLE_SIZE, size must be greater than 0

If size is not equal to WHOLE_SIZE, size must be less than or equal to than the size of Buffer minus offset

If Buffer was created with a sharing mode of SHARING_MODE_CONCURRENT, at least one of srcQueueFamilyIndex and dstQueueFamilyIndex must be QUEUE_FAMILY_IGNORED

If Buffer was created with a sharing mode of SHARING_MODE_CONCURRENT, and one of srcQueueFamilyIndex and dstQueueFamilyIndex is QUEUE_FAMILY_IGNORED, the other must be QUEUE_FAMILY_IGNORED or a special queue family reserved for external memory ownership transfers, as described in https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#synchronization-queue-transfers.

If Buffer was created with a sharing mode of SHARING_MODE_EXCLUSIVE and srcQueueFamilyIndex is QUEUE_FAMILY_IGNORED, dstQueueFamilyIndex must also be QUEUE_FAMILY_IGNORED

If Buffer was created with a sharing mode of SHARING_MODE_EXCLUSIVE and srcQueueFamilyIndex is not QUEUE_FAMILY_IGNORED, it must be a valid queue family or a special queue family reserved for external memory transfers, as described in https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#synchronization-queue-transfers.

If Buffer was created with a sharing mode of SHARING_MODE_EXCLUSIVE and dstQueueFamilyIndex is not QUEUE_FAMILY_IGNORED, it must be a valid queue family or a special queue family reserved for external memory transfers, as described in https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#synchronization-queue-transfers.

If Buffer was created with a sharing mode of SHARING_MODE_EXCLUSIVE, and srcQueueFamilyIndex and dstQueueFamilyIndex are not QUEUE_FAMILY_IGNORED, at least one of them must be the same as the family of the queue that will execute this barrier

If Buffer is non-sparse then it must be bound completely and contiguously to a single DeviceMemory object

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_BUFFER_MEMORY_BARRIER

pNext must be NULL

Buffer must be a valid Buffer handle

See Also

AccessFlags, Buffer, DeviceSize, StructureType, cmdPipelineBarrier, cmdWaitEvents

Constructors

BufferMemoryBarrier

Fields

srcAccessMask :: AccessFlags

srcAccessMask is a bitmask of AccessFlagBits specifying a source access mask.

dstAccessMask :: AccessFlags

dstAccessMask is a bitmask of AccessFlagBits specifying a destination access mask.

srcQueueFamilyIndex :: Word32

srcQueueFamilyIndex is the source queue family for a queue family ownership transfer.

dstQueueFamilyIndex :: Word32

dstQueueFamilyIndex is the destination queue family for a queue family ownership transfer.

buffer :: Buffer

Buffer is a handle to the buffer whose backing memory is affected by the barrier.

offset :: DeviceSize

offset is an offset in bytes into the backing memory for Buffer; this is relative to the base offset as bound to the buffer (see bindBufferMemory).

size :: DeviceSize

size is a size in bytes of the affected area of backing memory for Buffer, or WHOLE_SIZE to use the range from offset to the end of the buffer.

data ImageMemoryBarrier (es :: [Type])

VkImageMemoryBarrier - Structure specifying the parameters of an image memory barrier

Description

The first access scope is limited to access to memory through the specified image subresource range, via access types in the source access mask specified by srcAccessMask. If srcAccessMask includes ACCESS_HOST_WRITE_BIT, memory writes performed by that access type are also made visible, as that access type is not performed through a resource.

The second access scope is limited to access to memory through the specified image subresource range, via access types in the destination access mask specified by dstAccessMask. If dstAccessMask includes ACCESS_HOST_WRITE_BIT or ACCESS_HOST_READ_BIT, available memory writes are also made visible to accesses of those types, as those access types are not performed through a resource.

If srcQueueFamilyIndex is not equal to dstQueueFamilyIndex, and srcQueueFamilyIndex is equal to the current queue family, then the memory barrier defines a queue family release operation for the specified image subresource range, and the second access scope includes no access, as if dstAccessMask was 0.

If dstQueueFamilyIndex is not equal to srcQueueFamilyIndex, and dstQueueFamilyIndex is equal to the current queue family, then the memory barrier defines a queue family acquire operation for the specified image subresource range, and the first access scope includes no access, as if srcAccessMask was 0.

If oldLayout is not equal to newLayout, then the memory barrier defines an image layout transition for the specified image subresource range.

Layout transitions that are performed via image memory barriers execute in their entirety in submission order, relative to other image layout transitions submitted to the same queue, including those performed by render passes. In effect there is an implicit execution dependency from each such layout transition to all layout transitions previously submitted to the same queue.

The image layout of each image subresource of a depth/stencil image created with IMAGE_CREATE_SAMPLE_LOCATIONS_COMPATIBLE_DEPTH_BIT_EXT is dependent on the last sample locations used to render to the image subresource as a depth/stencil attachment, thus when the Image member of a ImageMemoryBarrier is an image created with this flag the application can include a SampleLocationsInfoEXT structure in the pNext chain of ImageMemoryBarrier to specify the sample locations to use during the image layout transition.

If the SampleLocationsInfoEXT structure included in the pNext chain of ImageMemoryBarrier does not match the sample location state last used to render to the image subresource range specified by subresourceRange or if no SampleLocationsInfoEXT structure is included in the pNext chain of ImageMemoryBarrier, then the contents of the given image subresource range becomes undefined as if oldLayout would equal IMAGE_LAYOUT_UNDEFINED.

If Image has a multi-planar format and the image is disjoint, then including IMAGE_ASPECT_COLOR_BIT in the aspectMask member of subresourceRange is equivalent to including IMAGE_ASPECT_PLANE_0_BIT, IMAGE_ASPECT_PLANE_1_BIT, and (for three-plane formats only) IMAGE_ASPECT_PLANE_2_BIT.

Valid Usage

oldLayout must be IMAGE_LAYOUT_UNDEFINED or the current layout of the image subresources affected by the barrier

newLayout must not be IMAGE_LAYOUT_UNDEFINED or IMAGE_LAYOUT_PREINITIALIZED

If Image was created with a sharing mode of SHARING_MODE_CONCURRENT, at least one of srcQueueFamilyIndex and dstQueueFamilyIndex must be QUEUE_FAMILY_IGNORED

If Image was created with a sharing mode of SHARING_MODE_CONCURRENT, and one of srcQueueFamilyIndex and dstQueueFamilyIndex is QUEUE_FAMILY_IGNORED, the other must be QUEUE_FAMILY_IGNORED or a special queue family reserved for external memory transfers, as described in https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#synchronization-queue-transfers.

If Image was created with a sharing mode of SHARING_MODE_EXCLUSIVE and srcQueueFamilyIndex is QUEUE_FAMILY_IGNORED, dstQueueFamilyIndex must also be QUEUE_FAMILY_IGNORED.

If Image was created with a sharing mode of SHARING_MODE_EXCLUSIVE and srcQueueFamilyIndex is not QUEUE_FAMILY_IGNORED, it must be a valid queue family or a special queue family reserved for external memory transfers, as described in https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#synchronization-queue-transfers.

If Image was created with a sharing mode of SHARING_MODE_EXCLUSIVE and dstQueueFamilyIndex is not QUEUE_FAMILY_IGNORED, it must be a valid queue family or a special queue family reserved for external memory transfers, as described in https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#synchronization-queue-transfers.

If Image was created with a sharing mode of SHARING_MODE_EXCLUSIVE, and srcQueueFamilyIndex and dstQueueFamilyIndex are not QUEUE_FAMILY_IGNORED, at least one of them must be the same as the family of the queue that will execute this barrier

subresourceRange.baseMipLevel must be less than the mipLevels specified in ImageCreateInfo when Image was created

If subresourceRange.levelCount is not REMAINING_MIP_LEVELS, subresourceRange.baseMipLevel + subresourceRange.levelCount must be less than or equal to the mipLevels specified in ImageCreateInfo when Image was created

subresourceRange.baseArrayLayer must be less than the arrayLayers specified in ImageCreateInfo when Image was created

If subresourceRange.layerCount is not REMAINING_ARRAY_LAYERS, subresourceRange.baseArrayLayer + subresourceRange.layerCount must be less than or equal to the arrayLayers specified in ImageCreateInfo when Image was created

If Image has a depth/stencil format with both depth and stencil and the separateDepthStencilLayouts feature is enabled, then the aspectMask member of subresourceRange must include either or both IMAGE_ASPECT_DEPTH_BIT and IMAGE_ASPECT_STENCIL_BIT

If Image has a depth/stencil format with both depth and stencil and the separateDepthStencilLayouts feature is not enabled, then the aspectMask member of subresourceRange must include both IMAGE_ASPECT_DEPTH_BIT and IMAGE_ASPECT_STENCIL_BIT

If Image has a single-plane color format or is not disjoint, then the aspectMask member of subresourceRange must be IMAGE_ASPECT_COLOR_BIT

If Image has a multi-planar format and the image is disjoint, then the aspectMask member of subresourceRange must include either at least one of IMAGE_ASPECT_PLANE_0_BIT, IMAGE_ASPECT_PLANE_1_BIT, and IMAGE_ASPECT_PLANE_2_BIT; or must include IMAGE_ASPECT_COLOR_BIT

If Image has a multi-planar format with only two planes, then the aspectMask member of subresourceRange must not include IMAGE_ASPECT_PLANE_2_BIT

If either oldLayout or newLayout is IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL then Image must have been created with IMAGE_USAGE_COLOR_ATTACHMENT_BIT set

If either oldLayout or newLayout is IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL then Image must have been created with IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT set

If either oldLayout or newLayout is IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL then Image must have been created with IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT set

If either oldLayout or newLayout is IMAGE_LAYOUT_DEPTH_READ_ONLY_STENCIL_ATTACHMENT_OPTIMAL then Image must have been created with IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT set

If either oldLayout or newLayout is IMAGE_LAYOUT_DEPTH_ATTACHMENT_STENCIL_READ_ONLY_OPTIMAL then Image must have been created with IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT set

If either oldLayout or newLayout is IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL then Image must have been created with IMAGE_USAGE_SAMPLED_BIT or IMAGE_USAGE_INPUT_ATTACHMENT_BIT set

If either oldLayout or newLayout is IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL then Image must have been created with IMAGE_USAGE_TRANSFER_SRC_BIT set

If either oldLayout or newLayout is IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL then Image must have been created with IMAGE_USAGE_TRANSFER_DST_BIT set

If Image is non-sparse then it must be bound completely and contiguously to a single DeviceMemory object

If either oldLayout or newLayout is IMAGE_LAYOUT_SHADING_RATE_OPTIMAL_NV then Image must have been created with IMAGE_USAGE_SHADING_RATE_IMAGE_BIT_NV set

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER

pNext must be NULL or a pointer to a valid instance of SampleLocationsInfoEXT

The sType value of each struct in the pNext chain must be unique

oldLayout must be a valid ImageLayout value

newLayout must be a valid ImageLayout value

Image must be a valid Image handle

subresourceRange must be a valid ImageSubresourceRange structure

See Also

AccessFlags, Image, ImageLayout, ImageSubresourceRange, StructureType, cmdPipelineBarrier, cmdWaitEvents

Constructors

ImageMemoryBarrier

Fields

next :: Chain es

pNext is NULL or a pointer to an extension-specific structure.

srcAccessMask :: AccessFlags

srcAccessMask is a bitmask of AccessFlagBits specifying a source access mask.

dstAccessMask :: AccessFlags

dstAccessMask is a bitmask of AccessFlagBits specifying a destination access mask.

oldLayout :: ImageLayout

oldLayout is the old layout in an image layout transition.

newLayout :: ImageLayout

newLayout is the new layout in an image layout transition.

srcQueueFamilyIndex :: Word32

srcQueueFamilyIndex is the source queue family for a queue family ownership transfer.

dstQueueFamilyIndex :: Word32

dstQueueFamilyIndex is the destination queue family for a queue family ownership transfer.

image :: Image

Image is a handle to the image affected by this barrier.

subresourceRange :: ImageSubresourceRange

subresourceRange describes the image subresource range within Image that is affected by this barrier.

data DrawIndirectCommand

VkDrawIndirectCommand - Structure specifying a draw indirect command

Description

The members of DrawIndirectCommand have the same meaning as the similarly named parameters of cmdDraw.

Valid Usage

For a given vertex buffer binding, any attribute data fetched must be entirely contained within the corresponding vertex buffer binding, as described in https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#fxvertex-input

If the drawIndirectFirstInstance feature is not enabled, firstInstance must be 0

See Also

cmdDrawIndirect

Constructors

DrawIndirectCommand

Fields

vertexCount :: Word32

vertexCount is the number of vertices to draw.

instanceCount :: Word32

instanceCount is the number of instances to draw.

firstVertex :: Word32

firstVertex is the index of the first vertex to draw.

firstInstance :: Word32

firstInstance is the instance ID of the first instance to draw.

data DrawIndexedIndirectCommand

VkDrawIndexedIndirectCommand - Structure specifying a draw indexed indirect command

Description

The members of DrawIndexedIndirectCommand have the same meaning as the similarly named parameters of cmdDrawIndexed.

Valid Usage

For a given vertex buffer binding, any attribute data fetched must be entirely contained within the corresponding vertex buffer binding, as described in https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#fxvertex-input

(indexSize * (firstIndex + indexCount) + offset) must be less than or equal to the size of the bound index buffer, with indexSize being based on the type specified by IndexType, where the index buffer, IndexType, and offset are specified via cmdBindIndexBuffer

If the drawIndirectFirstInstance feature is not enabled, firstInstance must be 0

See Also

cmdDrawIndexedIndirect

Constructors

DrawIndexedIndirectCommand

Fields

indexCount :: Word32

indexCount is the number of vertices to draw.

instanceCount :: Word32

instanceCount is the number of instances to draw.

firstIndex :: Word32

firstIndex is the base index within the index buffer.

vertexOffset :: Int32

vertexOffset is the value added to the vertex index before indexing into the vertex buffer.

firstInstance :: Word32

firstInstance is the instance ID of the first instance to draw.

data DispatchIndirectCommand

VkDispatchIndirectCommand - Structure specifying a dispatch indirect command

Description

The members of DispatchIndirectCommand have the same meaning as the corresponding parameters of cmdDispatch.

Valid Usage

See Also

cmdDispatchIndirect

Constructors

DispatchIndirectCommand

Fields

x :: Word32

x must be less than or equal to PhysicalDeviceLimits::maxComputeWorkGroupCount[0]

y :: Word32

y must be less than or equal to PhysicalDeviceLimits::maxComputeWorkGroupCount[1]

z :: Word32

z must be less than or equal to PhysicalDeviceLimits::maxComputeWorkGroupCount[2]

data BaseOutStructure

VkBaseOutStructure - Base structure for a read-only pointer chain

Description

BaseOutStructure can be used to facilitate iterating through a structure pointer chain that returns data back to the application.

See Also

BaseOutStructure, StructureType

Constructors

BaseOutStructure

Fields

sType :: StructureType

sType is the structure type of the structure being iterated through.

next :: Ptr BaseOutStructure

pNext is NULL or a pointer to the next structure in a structure chain.

data BaseInStructure

VkBaseInStructure - Base structure for a read-only pointer chain

Description

BaseInStructure can be used to facilitate iterating through a read-only structure pointer chain.

See Also

BaseInStructure, StructureType

Constructors

BaseInStructure

Fields

sType :: StructureType

sType is the structure type of the structure being iterated through.

next :: Ptr BaseInStructure

pNext is NULL or a pointer to the next structure in a structure chain.

newtype ObjectType

VkObjectType - Specify an enumeration to track object handle types

Description

'

ObjectType 	 Vulkan Handle Type

OBJECT_TYPE_UNKNOWN 	 Unknown/Undefined Handle

OBJECT_TYPE_INSTANCE 	 Instance

OBJECT_TYPE_PHYSICAL_DEVICE 	 PhysicalDevice

OBJECT_TYPE_DEVICE 	 Device

OBJECT_TYPE_QUEUE 	 Queue

OBJECT_TYPE_SEMAPHORE 	 Semaphore

OBJECT_TYPE_COMMAND_BUFFER 	 CommandBuffer

OBJECT_TYPE_FENCE 	 Fence

OBJECT_TYPE_DEVICE_MEMORY 	 DeviceMemory

OBJECT_TYPE_BUFFER 	 Buffer

OBJECT_TYPE_IMAGE 	 Image

OBJECT_TYPE_EVENT 	 Event

OBJECT_TYPE_QUERY_POOL 	 QueryPool

OBJECT_TYPE_BUFFER_VIEW 	 BufferView

OBJECT_TYPE_IMAGE_VIEW 	 ImageView

OBJECT_TYPE_SHADER_MODULE 	 ShaderModule

OBJECT_TYPE_PIPELINE_CACHE 	 PipelineCache

OBJECT_TYPE_PIPELINE_LAYOUT 	 PipelineLayout

OBJECT_TYPE_RENDER_PASS 	 RenderPass

OBJECT_TYPE_PIPELINE 	 Pipeline

OBJECT_TYPE_DESCRIPTOR_SET_LAYOUT 	 DescriptorSetLayout

OBJECT_TYPE_SAMPLER 	 Sampler

OBJECT_TYPE_DESCRIPTOR_POOL 	 DescriptorPool

OBJECT_TYPE_DESCRIPTOR_SET 	 DescriptorSet

OBJECT_TYPE_FRAMEBUFFER 	 Framebuffer

OBJECT_TYPE_COMMAND_POOL 	 CommandPool

OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION 	 SamplerYcbcrConversion

OBJECT_TYPE_DESCRIPTOR_UPDATE_TEMPLATE 	 DescriptorUpdateTemplate

OBJECT_TYPE_SURFACE_KHR 	 SurfaceKHR

OBJECT_TYPE_SWAPCHAIN_KHR 	 SwapchainKHR

OBJECT_TYPE_DISPLAY_KHR 	 DisplayKHR

OBJECT_TYPE_DISPLAY_MODE_KHR 	 DisplayModeKHR

OBJECT_TYPE_DEBUG_REPORT_CALLBACK_EXT 	 DebugReportCallbackEXT

OBJECT_TYPE_OBJECT_TABLE_NVX 	 ObjectTableNVX

OBJECT_TYPE_INDIRECT_COMMANDS_LAYOUT_NVX 	 IndirectCommandsLayoutNVX

OBJECT_TYPE_DEBUG_UTILS_MESSENGER_EXT 	 DebugUtilsMessengerEXT

OBJECT_TYPE_VALIDATION_CACHE_EXT 	 ValidationCacheEXT

OBJECT_TYPE_ACCELERATION_STRUCTURE_NV 	 AccelerationStructureNV

OBJECT_TYPE_PERFORMANCE_CONFIGURATION_INTEL 	 PerformanceConfigurationINTEL

VkObjectType and Vulkan Handle Relationship

See Also

DebugUtilsObjectNameInfoEXT, DebugUtilsObjectTagInfoEXT

Constructors

ObjectType Int32

Bundled Patterns

pattern OBJECT_TYPE_UNKNOWN :: ObjectType

pattern OBJECT_TYPE_INSTANCE :: ObjectType

pattern OBJECT_TYPE_PHYSICAL_DEVICE :: ObjectType

pattern OBJECT_TYPE_DEVICE :: ObjectType

pattern OBJECT_TYPE_QUEUE :: ObjectType

pattern OBJECT_TYPE_SEMAPHORE :: ObjectType

pattern OBJECT_TYPE_COMMAND_BUFFER :: ObjectType

pattern OBJECT_TYPE_FENCE :: ObjectType

pattern OBJECT_TYPE_DEVICE_MEMORY :: ObjectType

pattern OBJECT_TYPE_BUFFER :: ObjectType

pattern OBJECT_TYPE_IMAGE :: ObjectType

pattern OBJECT_TYPE_EVENT :: ObjectType

pattern OBJECT_TYPE_QUERY_POOL :: ObjectType

pattern OBJECT_TYPE_BUFFER_VIEW :: ObjectType

pattern OBJECT_TYPE_IMAGE_VIEW :: ObjectType

pattern OBJECT_TYPE_SHADER_MODULE :: ObjectType

pattern OBJECT_TYPE_PIPELINE_CACHE :: ObjectType

pattern OBJECT_TYPE_PIPELINE_LAYOUT :: ObjectType

pattern OBJECT_TYPE_RENDER_PASS :: ObjectType

pattern OBJECT_TYPE_PIPELINE :: ObjectType

pattern OBJECT_TYPE_DESCRIPTOR_SET_LAYOUT :: ObjectType

pattern OBJECT_TYPE_SAMPLER :: ObjectType

pattern OBJECT_TYPE_DESCRIPTOR_POOL :: ObjectType

pattern OBJECT_TYPE_DESCRIPTOR_SET :: ObjectType

pattern OBJECT_TYPE_FRAMEBUFFER :: ObjectType

pattern OBJECT_TYPE_COMMAND_POOL :: ObjectType

pattern OBJECT_TYPE_PERFORMANCE_CONFIGURATION_INTEL :: ObjectType

pattern OBJECT_TYPE_ACCELERATION_STRUCTURE_NV :: ObjectType

pattern OBJECT_TYPE_VALIDATION_CACHE_EXT :: ObjectType

pattern OBJECT_TYPE_DEBUG_UTILS_MESSENGER_EXT :: ObjectType

pattern OBJECT_TYPE_INDIRECT_COMMANDS_LAYOUT_NVX :: ObjectType

pattern OBJECT_TYPE_OBJECT_TABLE_NVX :: ObjectType

pattern OBJECT_TYPE_DEBUG_REPORT_CALLBACK_EXT :: ObjectType

pattern OBJECT_TYPE_DISPLAY_MODE_KHR :: ObjectType

pattern OBJECT_TYPE_DISPLAY_KHR :: ObjectType

pattern OBJECT_TYPE_SWAPCHAIN_KHR :: ObjectType

pattern OBJECT_TYPE_SURFACE_KHR :: ObjectType

pattern OBJECT_TYPE_DESCRIPTOR_UPDATE_TEMPLATE :: ObjectType

pattern OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION :: ObjectType

newtype VendorId

VkVendorId - Khronos vendor IDs

Description

Note

Khronos vendor IDs may be allocated by vendors at any time. Only the latest canonical versions of this Specification, of the corresponding vk.xml API Registry, and of the corresponding vulkan_core.h header file must contain all reserved Khronos vendor IDs.

Only Khronos vendor IDs are given symbolic names at present. PCI vendor IDs returned by the implementation can be looked up in the PCI-SIG database.

See Also

No cross-references are available

Constructors

VendorId Int32

Bundled Patterns

pattern VENDOR_ID_VIV :: VendorId

pattern VENDOR_ID_VSI :: VendorId

pattern VENDOR_ID_KAZAN :: VendorId

pattern VENDOR_ID_CODEPLAY :: VendorId