Graphics.Vulkan.Core10.PipelineLayout

Documentation

type DescriptorSetLayout = VkDescriptorSetLayout

VkDescriptorSetLayout - Opaque handle to a descriptor set layout object

See Also

VkDescriptorSetAllocateInfo, VkDescriptorUpdateTemplateCreateInfo, VkPipelineLayoutCreateInfo, vkCreateDescriptorSetLayout, vkDestroyDescriptorSetLayout

type PipelineLayoutCreateFlags = VkPipelineLayoutCreateFlags

VkPipelineLayoutCreateFlags - Reserved for future use

Description

VkPipelineLayoutCreateFlags is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkPipelineLayoutCreateInfo

withCStructPipelineLayoutCreateInfo :: PipelineLayoutCreateInfo -> (VkPipelineLayoutCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkPipelineLayoutCreateInfo and marshal a PipelineLayoutCreateInfo into it. The VkPipelineLayoutCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructPipelineLayoutCreateInfo :: VkPipelineLayoutCreateInfo -> IO PipelineLayoutCreateInfo

A function to read a VkPipelineLayoutCreateInfo and all additional structures in the pointer chain into a PipelineLayoutCreateInfo.

data PipelineLayoutCreateInfo

VkPipelineLayoutCreateInfo - Structure specifying the parameters of a newly created pipeline layout object

Valid Usage

setLayoutCount must be less than or equal to VkPhysicalDeviceLimits::maxBoundDescriptorSets

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_SAMPLER and VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER accessible to any shader stage across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxPerStageDescriptorSamplers

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER and VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC accessible to any shader stage across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxPerStageDescriptorUniformBuffers

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_STORAGE_BUFFER and VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC accessible to any shader stage across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxPerStageDescriptorStorageBuffers

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE, and VK_DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER accessible to any shader stage across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxPerStageDescriptorSampledImages

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, and VK_DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER accessible to any shader stage across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxPerStageDescriptorStorageImages

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT accessible to any given shader stage across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxPerStageDescriptorInputAttachments

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_SAMPLER and VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxDescriptorSetSamplers

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxDescriptorSetUniformBuffers

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxDescriptorSetUniformBuffersDynamic

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_STORAGE_BUFFER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxDescriptorSetStorageBuffers

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxDescriptorSetStorageBuffersDynamic

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE, and VK_DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxDescriptorSetSampledImages

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, and VK_DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxDescriptorSetStorageImages

The total number of descriptors of the type VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to VkPhysicalDeviceLimits::maxDescriptorSetInputAttachments

Any two elements of pPushConstantRanges must not include the same stage in stageFlags

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO

pNext must be NULL

flags must be 0

If setLayoutCount is not 0, pSetLayouts must be a valid pointer to an array of setLayoutCount valid VkDescriptorSetLayout handles

If pushConstantRangeCount is not 0, pPushConstantRanges must be a valid pointer to an array of pushConstantRangeCount valid VkPushConstantRange structures

See Also

VkDescriptorSetLayout, VkPipelineLayoutCreateFlags, VkPushConstantRange, VkStructureType, vkCreatePipelineLayout

Constructors

PipelineLayoutCreateInfo

Fields

next :: Maybe SomeVkStruct

flags :: PipelineLayoutCreateFlags

setLayouts :: Vector DescriptorSetLayout

pushConstantRanges :: Vector PushConstantRange

withCStructPushConstantRange :: PushConstantRange -> (VkPushConstantRange -> IO a) -> IO a

A function to temporarily allocate memory for a VkPushConstantRange and marshal a PushConstantRange into it. The VkPushConstantRange is only valid inside the provided computation and must not be returned out of it.

fromCStructPushConstantRange :: VkPushConstantRange -> IO PushConstantRange

A function to read a VkPushConstantRange and all additional structures in the pointer chain into a PushConstantRange.

data PushConstantRange

VkPushConstantRange - Structure specifying a push constant range

Valid Usage (Implicit)

See Also

VkPipelineLayoutCreateInfo, VkShaderStageFlags

Constructors

PushConstantRange

Fields

stageFlags :: ShaderStageFlags

offset :: Word32

size :: Word32

type ShaderStageFlags = ShaderStageFlagBits

VkShaderStageFlags - Bitmask of VkShaderStageFlagBits

Description

VkShaderStageFlags is a bitmask type for setting a mask of zero or more VkShaderStageFlagBits.

See Also

VkDescriptorSetLayoutBinding, VkObjectTablePushConstantEntryNVX, VkPhysicalDeviceCooperativeMatrixPropertiesNV, VkPhysicalDeviceSubgroupProperties, VkPushConstantRange, VkShaderStageFlagBits, VkShaderStatisticsInfoAMD, vkCmdPushConstants

createPipelineLayout :: Device -> PipelineLayoutCreateInfo -> Maybe AllocationCallbacks -> IO PipelineLayout

vkCreatePipelineLayout - Creates a new pipeline layout object

Parameters

device is the logical device that creates the pipeline layout.

pCreateInfo is a pointer to an instance of the VkPipelineLayoutCreateInfo structure specifying the state of the pipeline layout object.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pPipelineLayout points to a VkPipelineLayout handle in which the resulting pipeline layout object is returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkPipelineLayoutCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pPipelineLayout must be a valid pointer to a VkPipelineLayout handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDevice, VkPipelineLayout, VkPipelineLayoutCreateInfo

destroyPipelineLayout :: Device -> PipelineLayout -> Maybe AllocationCallbacks -> IO ()

vkDestroyPipelineLayout - Destroy a pipeline layout object

Parameters

device is the logical device that destroys the pipeline layout.

pipelineLayout is the pipeline layout to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

If VkAllocationCallbacks were provided when pipelineLayout was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when pipelineLayout was created, pAllocator must be NULL

pipelineLayout must not have been passed to any vkCmd* command for any command buffers that are still in the recording state when vkDestroyPipelineLayout is called

Valid Usage (Implicit)

device must be a valid VkDevice handle

If pipelineLayout is not VK_NULL_HANDLE, pipelineLayout must be a valid VkPipelineLayout handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

If pipelineLayout is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to pipelineLayout must be externally synchronized

See Also

VkAllocationCallbacks, VkDevice, VkPipelineLayout

withPipelineLayout :: Device -> PipelineLayoutCreateInfo -> Maybe AllocationCallbacks -> (PipelineLayout -> IO a) -> IO a

A safe wrapper for createPipelineLayout and destroyPipelineLayout using bracket

The allocated value must not be returned from the provided computation