Graphics.Vulkan.Core10.QueueSemaphore

Documentation

type SemaphoreCreateFlags = VkSemaphoreCreateFlags

VkSemaphoreCreateFlags - Reserved for future use

Description

VkSemaphoreCreateFlags is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkSemaphoreCreateInfo

withCStructSemaphoreCreateInfo :: SemaphoreCreateInfo -> (VkSemaphoreCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkSemaphoreCreateInfo and marshal a SemaphoreCreateInfo into it. The VkSemaphoreCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructSemaphoreCreateInfo :: VkSemaphoreCreateInfo -> IO SemaphoreCreateInfo

A function to read a VkSemaphoreCreateInfo and all additional structures in the pointer chain into a SemaphoreCreateInfo.

data SemaphoreCreateInfo

VkSemaphoreCreateInfo - Structure specifying parameters of a newly created semaphore

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_SEMAPHORE_CREATE_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkExportSemaphoreCreateInfo or VkExportSemaphoreWin32HandleInfoKHR

Each sType member in the pNext chain must be unique

flags must be 0

See Also

VkSemaphoreCreateFlags, VkStructureType, vkCreateSemaphore

Constructors

SemaphoreCreateInfo

Fields

next :: Maybe SomeVkStruct

flags :: SemaphoreCreateFlags

createSemaphore :: Device -> SemaphoreCreateInfo -> Maybe AllocationCallbacks -> IO Semaphore

vkCreateSemaphore - Create a new queue semaphore object

Parameters

device is the logical device that creates the semaphore.

pCreateInfo is a pointer to an instance of the VkSemaphoreCreateInfo structure which contains information about how the semaphore is to be created.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pSemaphore points to a handle in which the resulting semaphore object is returned.

Description

When created, the semaphore is in the unsignaled state.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkSemaphoreCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSemaphore must be a valid pointer to a VkSemaphore handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDevice, VkSemaphore, VkSemaphoreCreateInfo

destroySemaphore :: Device -> Semaphore -> Maybe AllocationCallbacks -> IO ()

vkDestroySemaphore - Destroy a semaphore object

Parameters

device is the logical device that destroys the semaphore.

semaphore is the handle of the semaphore to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted batches that refer to semaphore must have completed execution

If VkAllocationCallbacks were provided when semaphore was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when semaphore was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid VkDevice handle

If semaphore is not VK_NULL_HANDLE, semaphore must be a valid VkSemaphore handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

If semaphore is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to semaphore must be externally synchronized

See Also

VkAllocationCallbacks, VkDevice, VkSemaphore

withSemaphore :: Device -> SemaphoreCreateInfo -> Maybe AllocationCallbacks -> (Semaphore -> IO a) -> IO a

A safe wrapper for createSemaphore and destroySemaphore using bracket

The allocated value must not be returned from the provided computation