Graphics.Vulkan.Core10.Sampler

Documentation

type BorderColor = VkBorderColor

VkBorderColor - Specify border color used for texture lookups

Description

These colors are described in detail in Texel Replacement.

See Also

VkSamplerCreateInfo

pattern BORDER_COLOR_FLOAT_TRANSPARENT_BLACK :: a ~ BorderColor => a

VK_BORDER_COLOR_FLOAT_TRANSPARENT_BLACK specifies a transparent, floating-point format, black color.

pattern BORDER_COLOR_INT_TRANSPARENT_BLACK :: a ~ BorderColor => a

VK_BORDER_COLOR_INT_TRANSPARENT_BLACK specifies a transparent, integer format, black color.

pattern BORDER_COLOR_FLOAT_OPAQUE_BLACK :: a ~ BorderColor => a

VK_BORDER_COLOR_FLOAT_OPAQUE_BLACK specifies an opaque, floating-point format, black color.

pattern BORDER_COLOR_INT_OPAQUE_BLACK :: a ~ BorderColor => a

VK_BORDER_COLOR_INT_OPAQUE_BLACK specifies an opaque, integer format, black color.

pattern BORDER_COLOR_FLOAT_OPAQUE_WHITE :: a ~ BorderColor => a

VK_BORDER_COLOR_FLOAT_OPAQUE_WHITE specifies an opaque, floating-point format, white color.

pattern BORDER_COLOR_INT_OPAQUE_WHITE :: a ~ BorderColor => a

VK_BORDER_COLOR_INT_OPAQUE_WHITE specifies an opaque, integer format, white color.

type Filter = VkFilter

VkFilter - Specify filters used for texture lookups

Description

These filters are described in detail in Texel Filtering.

See Also

VkSamplerCreateInfo, VkSamplerYcbcrConversionCreateInfo, vkCmdBlitImage

pattern FILTER_NEAREST :: a ~ Filter => a

VK_FILTER_NEAREST specifies nearest filtering.

pattern FILTER_LINEAR :: a ~ Filter => a

VK_FILTER_LINEAR specifies linear filtering.

pattern FILTER_CUBIC_IMG :: a ~ Filter => a

type Sampler = VkSampler

VkSampler - Opaque handle to a sampler object

See Also

VkDescriptorImageInfo, VkDescriptorSetLayoutBinding, VkImageViewHandleInfoNVX, vkCreateSampler, vkDestroySampler

type SamplerAddressMode = VkSamplerAddressMode

VkSamplerAddressMode - Specify behavior of sampling with texture coordinates outside an image

See Also

VkSamplerCreateInfo

pattern SAMPLER_ADDRESS_MODE_REPEAT :: a ~ SamplerAddressMode => a

VK_SAMPLER_ADDRESS_MODE_REPEAT specifies that the repeat wrap mode will be used.

pattern SAMPLER_ADDRESS_MODE_MIRRORED_REPEAT :: a ~ SamplerAddressMode => a

VK_SAMPLER_ADDRESS_MODE_MIRRORED_REPEAT specifies that the mirrored repeat wrap mode will be used.

pattern SAMPLER_ADDRESS_MODE_CLAMP_TO_EDGE :: a ~ SamplerAddressMode => a

VK_SAMPLER_ADDRESS_MODE_CLAMP_TO_EDGE specifies that the clamp to edge wrap mode will be used.

pattern SAMPLER_ADDRESS_MODE_CLAMP_TO_BORDER :: a ~ SamplerAddressMode => a

VK_SAMPLER_ADDRESS_MODE_CLAMP_TO_BORDER specifies that the clamp to border wrap mode will be used.

pattern SAMPLER_ADDRESS_MODE_MIRROR_CLAMP_TO_EDGE :: a ~ SamplerAddressMode => a

VK_SAMPLER_ADDRESS_MODE_MIRROR_CLAMP_TO_EDGE specifies that the mirror clamp to edge wrap mode will be used. This is only valid if the https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#VK_KHR_sampler_mirror_clamp_to_edge extension is enabled.

type SamplerCreateFlagBits = VkSamplerCreateFlagBits

VkSamplerCreateFlagBits - Bitmask specifying additional parameters of sampler

See Also

VkSamplerCreateFlags

pattern SAMPLER_CREATE_SUBSAMPLED_BIT_EXT :: a ~ SamplerCreateFlagBits => a

pattern SAMPLER_CREATE_SUBSAMPLED_COARSE_RECONSTRUCTION_BIT_EXT :: a ~ SamplerCreateFlagBits => a

type SamplerCreateFlags = SamplerCreateFlagBits

VkSamplerCreateFlags - Reserved for future use

Description

VkSamplerCreateFlags is a bitmask type for setting a mask of zero or more VkSamplerCreateFlagBits.

See Also

VkSamplerCreateFlagBits, VkSamplerCreateInfo

withCStructSamplerCreateInfo :: SamplerCreateInfo -> (VkSamplerCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkSamplerCreateInfo and marshal a SamplerCreateInfo into it. The VkSamplerCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructSamplerCreateInfo :: VkSamplerCreateInfo -> IO SamplerCreateInfo

A function to read a VkSamplerCreateInfo and all additional structures in the pointer chain into a SamplerCreateInfo.

data SamplerCreateInfo

VkSamplerCreateInfo - Structure specifying parameters of a newly created sampler

Description

Note

magFilter values of VK_FILTER_NEAREST and VK_FILTER_LINEAR directly correspond to GL_NEAREST and GL_LINEAR magnification filters. minFilter and mipmapMode combine to correspond to the similarly named OpenGL minification filter of GL_minFilter_MIPMAP_mipmapMode (e.g. minFilter of VK_FILTER_LINEAR and mipmapMode of VK_SAMPLER_MIPMAP_MODE_NEAREST correspond to GL_LINEAR_MIPMAP_NEAREST).

There are no Vulkan filter modes that directly correspond to OpenGL minification filters of GL_LINEAR or GL_NEAREST, but they can be emulated using VK_SAMPLER_MIPMAP_MODE_NEAREST, minLod = 0, and maxLod = 0.25, and using minFilter = VK_FILTER_LINEAR or minFilter = VK_FILTER_NEAREST, respectively.

Note that using a maxLod of zero would cause magnification to always be performed, and the magFilter to always be used. This is valid, just not an exact match for OpenGL behavior. Clamping the maximum LOD to 0.25 allows the λ value to be non-zero and minification to be performed, while still always rounding down to the base level. If the minFilter and magFilter are equal, then using a maxLod of zero also works.

The maximum number of sampler objects which can be simultaneously created on a device is implementation-dependent and specified by the maxSamplerAllocationCount member of the VkPhysicalDeviceLimits structure. If maxSamplerAllocationCount is exceeded, vkCreateSampler will return VK_ERROR_TOO_MANY_OBJECTS.

Since VkSampler is a non-dispatchable handle type, implementations may return the same handle for sampler state vectors that are identical. In such cases, all such objects would only count once against the maxSamplerAllocationCount limit.

Valid Usage

The absolute value of mipLodBias must be less than or equal to VkPhysicalDeviceLimits::maxSamplerLodBias

maxLod must be greater than or equal to minLod

If the anisotropic sampling feature is not enabled, anisotropyEnable must be VK_FALSE

If anisotropyEnable is VK_TRUE, maxAnisotropy must be between 1.0 and VkPhysicalDeviceLimits::maxSamplerAnisotropy, inclusive

If unnormalizedCoordinates is VK_TRUE, minFilter and magFilter must be equal

If unnormalizedCoordinates is VK_TRUE, mipmapMode must be VK_SAMPLER_MIPMAP_MODE_NEAREST

If unnormalizedCoordinates is VK_TRUE, minLod and maxLod must be zero

If unnormalizedCoordinates is VK_TRUE, addressModeU and addressModeV must each be either VK_SAMPLER_ADDRESS_MODE_CLAMP_TO_EDGE or VK_SAMPLER_ADDRESS_MODE_CLAMP_TO_BORDER

If unnormalizedCoordinates is VK_TRUE, anisotropyEnable must be VK_FALSE

If unnormalizedCoordinates is VK_TRUE, compareEnable must be VK_FALSE

If any of addressModeU, addressModeV or addressModeW are VK_SAMPLER_ADDRESS_MODE_CLAMP_TO_BORDER, borderColor must be a valid VkBorderColor value

If the https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#VK_KHR_sampler_mirror_clamp_to_edge extension is not enabled, addressModeU, addressModeV and addressModeW must not be VK_SAMPLER_ADDRESS_MODE_MIRROR_CLAMP_TO_EDGE

If compareEnable is VK_TRUE, compareOp must be a valid VkCompareOp value

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_SAMPLER_CREATE_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkSamplerReductionModeCreateInfoEXT or VkSamplerYcbcrConversionInfo

Each sType member in the pNext chain must be unique

flags must be a valid combination of VkSamplerCreateFlagBits values

magFilter must be a valid VkFilter value

minFilter must be a valid VkFilter value

mipmapMode must be a valid VkSamplerMipmapMode value

addressModeU must be a valid VkSamplerAddressMode value

addressModeV must be a valid VkSamplerAddressMode value

addressModeW must be a valid VkSamplerAddressMode value

See Also

VkBool32, VkBorderColor, VkCompareOp, VkFilter, VkSamplerAddressMode, VkSamplerCreateFlags, VkSamplerMipmapMode, VkStructureType, vkCreateSampler

Constructors

SamplerCreateInfo

Fields

next :: Maybe SomeVkStruct

flags :: SamplerCreateFlags

magFilter :: Filter

minFilter :: Filter

mipmapMode :: SamplerMipmapMode

addressModeU :: SamplerAddressMode

addressModeV :: SamplerAddressMode

addressModeW :: SamplerAddressMode

mipLodBias :: CFloat

anisotropyEnable :: Bool

maxAnisotropy :: CFloat

compareEnable :: Bool

compareOp :: CompareOp

minLod :: CFloat

maxLod :: CFloat

borderColor :: BorderColor

unnormalizedCoordinates :: Bool

type SamplerMipmapMode = VkSamplerMipmapMode

VkSamplerMipmapMode - Specify mipmap mode used for texture lookups

Description

These modes are described in detail in Texel Filtering.

See Also

VkSamplerCreateInfo

pattern SAMPLER_MIPMAP_MODE_NEAREST :: a ~ SamplerMipmapMode => a

VK_SAMPLER_MIPMAP_MODE_NEAREST specifies nearest filtering.

pattern SAMPLER_MIPMAP_MODE_LINEAR :: a ~ SamplerMipmapMode => a

VK_SAMPLER_MIPMAP_MODE_LINEAR specifies linear filtering.

createSampler :: Device -> SamplerCreateInfo -> Maybe AllocationCallbacks -> IO Sampler

vkCreateSampler - Create a new sampler object

Parameters

device is the logical device that creates the sampler.

pCreateInfo is a pointer to an instance of the VkSamplerCreateInfo structure specifying the state of the sampler object.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pSampler points to a VkSampler handle in which the resulting sampler object is returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkSamplerCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSampler must be a valid pointer to a VkSampler handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_TOO_MANY_OBJECTS

See Also

VkAllocationCallbacks, VkDevice, VkSampler, VkSamplerCreateInfo

destroySampler :: Device -> Sampler -> Maybe AllocationCallbacks -> IO ()

vkDestroySampler - Destroy a sampler object

Parameters

device is the logical device that destroys the sampler.

sampler is the sampler to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to sampler must have completed execution

If VkAllocationCallbacks were provided when sampler was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when sampler was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid VkDevice handle

If sampler is not VK_NULL_HANDLE, sampler must be a valid VkSampler handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

If sampler is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to sampler must be externally synchronized

See Also

VkAllocationCallbacks, VkDevice, VkSampler

withSampler :: Device -> SamplerCreateInfo -> Maybe AllocationCallbacks -> (Sampler -> IO a) -> IO a

A safe wrapper for createSampler and destroySampler using bracket

The allocated value must not be returned from the provided computation