Graphics.Vulkan.Core10.Shader

Documentation

type ShaderModule = VkShaderModule

VkShaderModule - Opaque handle to a shader module object

See Also

VkPipelineShaderStageCreateInfo, vkCreateShaderModule, vkDestroyShaderModule

type ShaderModuleCreateFlags = VkShaderModuleCreateFlags

VkShaderModuleCreateFlags - Reserved for future use

Description

VkShaderModuleCreateFlags is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkShaderModuleCreateInfo

withCStructShaderModuleCreateInfo :: ShaderModuleCreateInfo -> (VkShaderModuleCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkShaderModuleCreateInfo and marshal a ShaderModuleCreateInfo into it. The VkShaderModuleCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructShaderModuleCreateInfo :: VkShaderModuleCreateInfo -> IO ShaderModuleCreateInfo

A function to read a VkShaderModuleCreateInfo and all additional structures in the pointer chain into a ShaderModuleCreateInfo.

data ShaderModuleCreateInfo

VkShaderModuleCreateInfo - Structure specifying parameters of a newly created shader module

Valid Usage

codeSize must be greater than 0

codeSize must be a multiple of 4

pCode must point to valid SPIR-V code, formatted and packed as described by the Khronos SPIR-V Specification

pCode must adhere to the validation rules described by the Validation Rules within a Module section of the SPIR-V Environment appendix

pCode must declare the Shader capability for SPIR-V code

pCode must not declare any capability that is not supported by the API, as described by the Capabilities section of the SPIR-V Environment appendix

If pCode declares any of the capabilities listed as optional in the SPIR-V Environment appendix, the corresponding feature(s) must be enabled.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO

pNext must be NULL or a pointer to a valid instance of VkShaderModuleValidationCacheCreateInfoEXT

flags must be 0

pCode must be a valid pointer to an array of \(\textrm{codeSize} \over 4\) uint32_t values

See Also

VkShaderModuleCreateFlags, VkStructureType, vkCreateShaderModule

Constructors

ShaderModuleCreateInfo

Fields

next :: Maybe SomeVkStruct

flags :: ShaderModuleCreateFlags

code :: Vector Word32

createShaderModule :: Device -> ShaderModuleCreateInfo -> Maybe AllocationCallbacks -> IO ShaderModule

vkCreateShaderModule - Creates a new shader module object

Parameters

device is the logical device that creates the shader module.

pCreateInfo is a pointer to an instance of the VkShaderModuleCreateInfo structure.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pShaderModule points to a VkShaderModule handle in which the resulting shader module object is returned.

Description

Once a shader module has been created, any entry points it contains can be used in pipeline shader stages as described in Compute Pipelines and Graphics Pipelines.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkShaderModuleCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pShaderModule must be a valid pointer to a VkShaderModule handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_INVALID_SHADER_NV

See Also

VkAllocationCallbacks, VkDevice, VkShaderModule, VkShaderModuleCreateInfo

destroyShaderModule :: Device -> ShaderModule -> Maybe AllocationCallbacks -> IO ()

vkDestroyShaderModule - Destroy a shader module

Parameters

device is the logical device that destroys the shader module.

shaderModule is the handle of the shader module to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Description

A shader module can be destroyed while pipelines created using its shaders are still in use.

Valid Usage

If VkAllocationCallbacks were provided when shaderModule was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when shaderModule was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid VkDevice handle

If shaderModule is not VK_NULL_HANDLE, shaderModule must be a valid VkShaderModule handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

If shaderModule is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to shaderModule must be externally synchronized

See Also

VkAllocationCallbacks, VkDevice, VkShaderModule

withShaderModule :: Device -> ShaderModuleCreateInfo -> Maybe AllocationCallbacks -> (ShaderModule -> IO a) -> IO a

A safe wrapper for createShaderModule and destroyShaderModule using bracket

The allocated value must not be returned from the provided computation