Graphics.Vulkan.Core10.SharedTypes

Documentation

data Offset2D

VkOffset2D - Structure specifying a two-dimensional offset

See Also

DisplayPlaneCapabilitiesKHR, Rect2D, RectLayerKHR

Constructors

Offset2D

Fields

x :: Int32

x is the x offset.

y :: Int32

y is the y offset.

data Offset3D

VkOffset3D - Structure specifying a three-dimensional offset

See Also

BufferImageCopy, ImageBlit, ImageCopy, ImageResolve, SparseImageMemoryBind

Constructors

Offset3D

Fields

x :: Int32

x is the x offset.

y :: Int32

y is the y offset.

z :: Int32

z is the z offset.

data Extent2D

VkExtent2D - Structure specifying a two-dimensional extent

See Also

DisplayModeParametersKHR, DisplayPlaneCapabilitiesKHR, DisplayPropertiesKHR, DisplaySurfaceCreateInfoKHR, MultisamplePropertiesEXT, PhysicalDeviceFragmentDensityMapPropertiesEXT, PhysicalDeviceSampleLocationsPropertiesEXT, PhysicalDeviceShadingRateImagePropertiesNV, Rect2D, RectLayerKHR, SampleLocationsInfoEXT, SurfaceCapabilities2EXT, SurfaceCapabilitiesKHR, SwapchainCreateInfoKHR, getRenderAreaGranularity

Constructors

Extent2D

Fields

width :: Word32

width is the width of the extent.

height :: Word32

height is the height of the extent.

data Extent3D

VkExtent3D - Structure specifying a three-dimensional extent

See Also

BufferImageCopy, ImageCopy, ImageCreateInfo, ImageFormatProperties, ImageResolve, QueueFamilyProperties, SparseImageFormatProperties, SparseImageMemoryBind

Constructors

Extent3D

Fields

width :: Word32

width is the width of the extent.

height :: Word32

height is the height of the extent.

depth :: Word32

depth is the depth of the extent.

data ImageSubresourceLayers

VkImageSubresourceLayers - Structure specifying an image subresource layers

Valid Usage

If aspectMask contains IMAGE_ASPECT_COLOR_BIT, it must not contain either of IMAGE_ASPECT_DEPTH_BIT or IMAGE_ASPECT_STENCIL_BIT

aspectMask must not contain IMAGE_ASPECT_METADATA_BIT

aspectMask must not include VK_IMAGE_ASPECT_MEMORY_PLANE_i_BIT_EXT for any index i.

layerCount must be greater than 0

Valid Usage (Implicit)

aspectMask must be a valid combination of ImageAspectFlagBits values

aspectMask must not be 0

See Also

BufferImageCopy, ImageAspectFlags, ImageBlit, ImageCopy, ImageResolve

Constructors

ImageSubresourceLayers

Fields

aspectMask :: ImageAspectFlags

aspectMask is a combination of ImageAspectFlagBits, selecting the color, depth and/or stencil aspects to be copied.

mipLevel :: Word32

mipLevel is the mipmap level to copy from.

baseArrayLayer :: Word32

baseArrayLayer and layerCount are the starting layer and number of layers to copy.

layerCount :: Word32

data ImageSubresourceRange

VkImageSubresourceRange - Structure specifying an image subresource range

Description

The number of mipmap levels and array layers must be a subset of the image subresources in the image. If an application wants to use all mip levels or layers in an image after the baseMipLevel or baseArrayLayer, it can set levelCount and layerCount to the special values REMAINING_MIP_LEVELS and REMAINING_ARRAY_LAYERS without knowing the exact number of mip levels or layers.

For cube and cube array image views, the layers of the image view starting at baseArrayLayer correspond to faces in the order +X, -X, +Y, -Y, +Z, -Z. For cube arrays, each set of six sequential layers is a single cube, so the number of cube maps in a cube map array view is layerCount / 6, and image array layer (baseArrayLayer + i) is face index (i mod 6) of cube i / 6. If the number of layers in the view, whether set explicitly in layerCount or implied by REMAINING_ARRAY_LAYERS, is not a multiple of 6, the last cube map in the array must not be accessed.

aspectMask must be only IMAGE_ASPECT_COLOR_BIT, IMAGE_ASPECT_DEPTH_BIT or IMAGE_ASPECT_STENCIL_BIT if Format is a color, depth-only or stencil-only format, respectively, except if Format is a multi-planar format. If using a depth/stencil format with both depth and stencil components, aspectMask must include at least one of IMAGE_ASPECT_DEPTH_BIT and IMAGE_ASPECT_STENCIL_BIT, and can include both.

When the ImageSubresourceRange structure is used to select a subset of the slices of a 3D image’s mip level in order to create a 2D or 2D array image view of a 3D image created with IMAGE_CREATE_2D_ARRAY_COMPATIBLE_BIT, baseArrayLayer and layerCount specify the first slice index and the number of slices to include in the created image view. Such an image view can be used as a framebuffer attachment that refers only to the specified range of slices of the selected mip level. However, any layout transitions performed on such an attachment view during a render pass instance still apply to the entire subresource referenced which includes all the slices of the selected mip level.

When using an image view of a depth/stencil image to populate a descriptor set (e.g. for sampling in the shader, or for use as an input attachment), the aspectMask must only include one bit and selects whether the image view is used for depth reads (i.e. using a floating-point sampler or input attachment in the shader) or stencil reads (i.e. using an unsigned integer sampler or input attachment in the shader). When an image view of a depth/stencil image is used as a depth/stencil framebuffer attachment, the aspectMask is ignored and both depth and stencil image subresources are used.

The ComponentMapping components member describes a remapping from components of the image to components of the vector returned by shader image instructions. This remapping must be identity for storage image descriptors, input attachment descriptors, framebuffer attachments, and any ImageView used with a combined image sampler that enables sampler Y′CBCR conversion.

When creating a ImageView, if sampler Y′CBCR conversion is enabled in the sampler, the aspectMask of a subresourceRange used by the ImageView must be IMAGE_ASPECT_COLOR_BIT.

When creating a ImageView, if sampler Y′CBCR conversion is not enabled in the sampler and the image Format is multi-planar, the image must have been created with IMAGE_CREATE_MUTABLE_FORMAT_BIT, and the aspectMask of the ImageView’s subresourceRange must be IMAGE_ASPECT_PLANE_0_BIT, IMAGE_ASPECT_PLANE_1_BIT or IMAGE_ASPECT_PLANE_2_BIT.

Valid Usage

If levelCount is not REMAINING_MIP_LEVELS, it must be greater than 0

If layerCount is not REMAINING_ARRAY_LAYERS, it must be greater than 0

If aspectMask includes IMAGE_ASPECT_COLOR_BIT, then it must not include any of IMAGE_ASPECT_PLANE_0_BIT, IMAGE_ASPECT_PLANE_1_BIT, or IMAGE_ASPECT_PLANE_2_BIT

aspectMask must not include VK_IMAGE_ASPECT_MEMORY_PLANE_i_BIT_EXT for any index i

Valid Usage (Implicit)

aspectMask must be a valid combination of ImageAspectFlagBits values

aspectMask must not be 0

See Also

ImageAspectFlags, ImageMemoryBarrier, ImageViewCreateInfo, cmdClearColorImage, cmdClearDepthStencilImage

Constructors

ImageSubresourceRange

Fields

aspectMask :: ImageAspectFlags

aspectMask is a bitmask of ImageAspectFlagBits specifying which aspect(s) of the image are included in the view.

baseMipLevel :: Word32

baseMipLevel is the first mipmap level accessible to the view.

levelCount :: Word32

levelCount is the number of mipmap levels (starting from baseMipLevel) accessible to the view.

baseArrayLayer :: Word32

baseArrayLayer is the first array layer accessible to the view.

layerCount :: Word32

layerCount is the number of array layers (starting from baseArrayLayer) accessible to the view.

data ClearDepthStencilValue

VkClearDepthStencilValue - Structure specifying a clear depth stencil value

Valid Usage

Unless the https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#VK_EXT_depth_range_unrestricted extension is enabled depth must be between 0.0 and 1.0, inclusive

See Also

ClearValue, cmdClearDepthStencilImage

Constructors

ClearDepthStencilValue

Fields

depth :: Float

depth is the clear value for the depth aspect of the depth/stencil attachment. It is a floating-point value which is automatically converted to the attachment’s format.

stencil :: Word32

stencil is the clear value for the stencil aspect of the depth/stencil attachment. It is a 32-bit integer value which is converted to the attachment’s format by taking the appropriate number of LSBs.

data ClearColorValue

Constructors

Float32 (Float, Float, Float, Float)

Int32 (Int32, Int32, Int32, Int32)

Uint32 (Word32, Word32, Word32, Word32)

data ClearValue

Constructors

Color ClearColorValue

DepthStencil ClearDepthStencilValue