Graphics.Vulkan.Core10.SparseResourceMemoryManagement

Documentation

withCStructBindSparseInfo :: BindSparseInfo -> (VkBindSparseInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkBindSparseInfo and marshal a BindSparseInfo into it. The VkBindSparseInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructBindSparseInfo :: VkBindSparseInfo -> IO BindSparseInfo

A function to read a VkBindSparseInfo and all additional structures in the pointer chain into a BindSparseInfo.

data BindSparseInfo

VkBindSparseInfo - Structure specifying a sparse binding operation

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_BIND_SPARSE_INFO

pNext must be NULL or a pointer to a valid instance of VkDeviceGroupBindSparseInfo

If waitSemaphoreCount is not 0, pWaitSemaphores must be a valid pointer to an array of waitSemaphoreCount valid VkSemaphore handles

If bufferBindCount is not 0, pBufferBinds must be a valid pointer to an array of bufferBindCount valid VkSparseBufferMemoryBindInfo structures

If imageOpaqueBindCount is not 0, pImageOpaqueBinds must be a valid pointer to an array of imageOpaqueBindCount valid VkSparseImageOpaqueMemoryBindInfo structures

If imageBindCount is not 0, pImageBinds must be a valid pointer to an array of imageBindCount valid VkSparseImageMemoryBindInfo structures

If signalSemaphoreCount is not 0, pSignalSemaphores must be a valid pointer to an array of signalSemaphoreCount valid VkSemaphore handles

Both of the elements of pSignalSemaphores, and the elements of pWaitSemaphores that are valid handles must have been created, allocated, or retrieved from the same VkDevice

See Also

VkSemaphore, VkSparseBufferMemoryBindInfo, VkSparseImageMemoryBindInfo, VkSparseImageOpaqueMemoryBindInfo, VkStructureType, vkQueueBindSparse

Constructors

BindSparseInfo

Fields

next :: Maybe SomeVkStruct

waitSemaphores :: Vector Semaphore

bufferBinds :: Vector SparseBufferMemoryBindInfo

imageOpaqueBinds :: Vector SparseImageOpaqueMemoryBindInfo

imageBinds :: Vector SparseImageMemoryBindInfo

signalSemaphores :: Vector Semaphore

type ImageAspectFlagBits = VkImageAspectFlagBits

VkImageAspectFlagBits - Bitmask specifying which aspects of an image are included in a view

See Also

VkBindImagePlaneMemoryInfo, VkImageAspectFlags, VkImagePlaneMemoryRequirementsInfo

pattern IMAGE_ASPECT_COLOR_BIT :: a ~ ImageAspectFlagBits => a

VK_IMAGE_ASPECT_COLOR_BIT specifies the color aspect.

pattern IMAGE_ASPECT_DEPTH_BIT :: a ~ ImageAspectFlagBits => a

VK_IMAGE_ASPECT_DEPTH_BIT specifies the depth aspect.

pattern IMAGE_ASPECT_STENCIL_BIT :: a ~ ImageAspectFlagBits => a

VK_IMAGE_ASPECT_STENCIL_BIT specifies the stencil aspect.

pattern IMAGE_ASPECT_METADATA_BIT :: a ~ ImageAspectFlagBits => a

VK_IMAGE_ASPECT_METADATA_BIT specifies the metadata aspect, used for sparse sparse resource operations.

pattern IMAGE_ASPECT_PLANE_0_BIT :: a ~ ImageAspectFlagBits => a

pattern IMAGE_ASPECT_PLANE_1_BIT :: a ~ ImageAspectFlagBits => a

pattern IMAGE_ASPECT_PLANE_2_BIT :: a ~ ImageAspectFlagBits => a

pattern IMAGE_ASPECT_MEMORY_PLANE_0_BIT_EXT :: a ~ ImageAspectFlagBits => a

pattern IMAGE_ASPECT_MEMORY_PLANE_1_BIT_EXT :: a ~ ImageAspectFlagBits => a

pattern IMAGE_ASPECT_MEMORY_PLANE_2_BIT_EXT :: a ~ ImageAspectFlagBits => a

pattern IMAGE_ASPECT_MEMORY_PLANE_3_BIT_EXT :: a ~ ImageAspectFlagBits => a

type ImageAspectFlags = ImageAspectFlagBits

VkImageAspectFlags - Bitmask of VkImageAspectFlagBits

Description

VkImageAspectFlags is a bitmask type for setting a mask of zero or more VkImageAspectFlagBits.

See Also

VkAttachmentReference2KHR, VkClearAttachment, VkImageAspectFlagBits, VkImageSubresource, VkImageSubresourceLayers, VkImageSubresourceRange, VkInputAttachmentAspectReference, VkSparseImageFormatProperties

withCStructImageSubresource :: ImageSubresource -> (VkImageSubresource -> IO a) -> IO a

A function to temporarily allocate memory for a VkImageSubresource and marshal a ImageSubresource into it. The VkImageSubresource is only valid inside the provided computation and must not be returned out of it.

fromCStructImageSubresource :: VkImageSubresource -> IO ImageSubresource

A function to read a VkImageSubresource and all additional structures in the pointer chain into a ImageSubresource.

data ImageSubresource

VkImageSubresource - Structure specifying an image subresource

Valid Usage (Implicit)

See Also

VkImageAspectFlags, VkSparseImageMemoryBind, vkGetImageSubresourceLayout

Constructors

ImageSubresource

Fields

aspectMask :: ImageAspectFlags

mipLevel :: Word32

arrayLayer :: Word32

withCStructOffset3D :: Offset3D -> (VkOffset3D -> IO a) -> IO a

A function to temporarily allocate memory for a VkOffset3D and marshal a Offset3D into it. The VkOffset3D is only valid inside the provided computation and must not be returned out of it.

fromCStructOffset3D :: VkOffset3D -> IO Offset3D

A function to read a VkOffset3D and all additional structures in the pointer chain into a Offset3D.

data Offset3D

VkOffset3D - Structure specifying a three-dimensional offset

See Also

VkBufferImageCopy, VkImageBlit, VkImageCopy, VkImageResolve, VkSparseImageMemoryBind

Constructors

Offset3D

Fields

x :: Int32

y :: Int32

z :: Int32

withCStructSparseBufferMemoryBindInfo :: SparseBufferMemoryBindInfo -> (VkSparseBufferMemoryBindInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkSparseBufferMemoryBindInfo and marshal a SparseBufferMemoryBindInfo into it. The VkSparseBufferMemoryBindInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructSparseBufferMemoryBindInfo :: VkSparseBufferMemoryBindInfo -> IO SparseBufferMemoryBindInfo

A function to read a VkSparseBufferMemoryBindInfo and all additional structures in the pointer chain into a SparseBufferMemoryBindInfo.

data SparseBufferMemoryBindInfo

VkSparseBufferMemoryBindInfo - Structure specifying a sparse buffer memory bind operation

Valid Usage (Implicit)

See Also

VkBindSparseInfo, VkBuffer, VkSparseMemoryBind

Constructors

SparseBufferMemoryBindInfo

Fields

buffer :: Buffer

binds :: Vector SparseMemoryBind

type SparseImageFormatFlagBits = VkSparseImageFormatFlagBits

VkSparseImageFormatFlagBits - Bitmask specifying additional information about a sparse image resource

See Also

VkSparseImageFormatFlags

pattern SPARSE_IMAGE_FORMAT_SINGLE_MIPTAIL_BIT :: a ~ SparseImageFormatFlagBits => a

VK_SPARSE_IMAGE_FORMAT_SINGLE_MIPTAIL_BIT specifies that the image uses a single mip tail region for all array layers.

pattern SPARSE_IMAGE_FORMAT_ALIGNED_MIP_SIZE_BIT :: a ~ SparseImageFormatFlagBits => a

VK_SPARSE_IMAGE_FORMAT_ALIGNED_MIP_SIZE_BIT specifies that the first mip level whose dimensions are not integer multiples of the corresponding dimensions of the sparse image block begins the mip tail region.

pattern SPARSE_IMAGE_FORMAT_NONSTANDARD_BLOCK_SIZE_BIT :: a ~ SparseImageFormatFlagBits => a

VK_SPARSE_IMAGE_FORMAT_NONSTANDARD_BLOCK_SIZE_BIT specifies that the image uses non-standard sparse image block dimensions, and the imageGranularity values do not match the standard sparse image block dimensions for the given format.

type SparseImageFormatFlags = SparseImageFormatFlagBits

VkSparseImageFormatFlags - Bitmask of VkSparseImageFormatFlagBits

Description

VkSparseImageFormatFlags is a bitmask type for setting a mask of zero or more VkSparseImageFormatFlagBits.

See Also

VkSparseImageFormatFlagBits, VkSparseImageFormatProperties

withCStructSparseImageFormatProperties :: SparseImageFormatProperties -> (VkSparseImageFormatProperties -> IO a) -> IO a

A function to temporarily allocate memory for a VkSparseImageFormatProperties and marshal a SparseImageFormatProperties into it. The VkSparseImageFormatProperties is only valid inside the provided computation and must not be returned out of it.

fromCStructSparseImageFormatProperties :: VkSparseImageFormatProperties -> IO SparseImageFormatProperties

A function to read a VkSparseImageFormatProperties and all additional structures in the pointer chain into a SparseImageFormatProperties.

data SparseImageFormatProperties

VkSparseImageFormatProperties - Structure specifying sparse image format properties

See Also

VkExtent3D, VkImageAspectFlags, VkSparseImageFormatFlags, VkSparseImageFormatProperties2, VkSparseImageMemoryRequirements, vkGetPhysicalDeviceSparseImageFormatProperties

Constructors

SparseImageFormatProperties

Fields

aspectMask :: ImageAspectFlags

imageGranularity :: Extent3D

flags :: SparseImageFormatFlags

withCStructSparseImageMemoryBind :: SparseImageMemoryBind -> (VkSparseImageMemoryBind -> IO a) -> IO a

A function to temporarily allocate memory for a VkSparseImageMemoryBind and marshal a SparseImageMemoryBind into it. The VkSparseImageMemoryBind is only valid inside the provided computation and must not be returned out of it.

fromCStructSparseImageMemoryBind :: VkSparseImageMemoryBind -> IO SparseImageMemoryBind

A function to read a VkSparseImageMemoryBind and all additional structures in the pointer chain into a SparseImageMemoryBind.

data SparseImageMemoryBind

VkSparseImageMemoryBind - Structure specifying sparse image memory bind

Valid Usage

If the sparse aliased residency feature is not enabled, and if any other resources are bound to ranges of memory, the range of memory being bound must not overlap with those bound ranges

memory and memoryOffset must match the memory requirements of the calling command’s image, as described in section https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#resources-association

subresource must be a valid image subresource for image (see https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#resources-image-views)

offset.x must be a multiple of the sparse image block width (VkSparseImageFormatProperties::imageGranularity.width) of the image

extent.width must either be a multiple of the sparse image block width of the image, or else (extent.width + offset.x) must equal the width of the image subresource

offset.y must be a multiple of the sparse image block height (VkSparseImageFormatProperties::imageGranularity.height) of the image

extent.height must either be a multiple of the sparse image block height of the image, or else (extent.height + offset.y) must equal the height of the image subresource

offset.z must be a multiple of the sparse image block depth (VkSparseImageFormatProperties::imageGranularity.depth) of the image

extent.depth must either be a multiple of the sparse image block depth of the image, or else (extent.depth + offset.z) must equal the depth of the image subresource

Valid Usage (Implicit)

subresource must be a valid VkImageSubresource structure

If memory is not VK_NULL_HANDLE, memory must be a valid VkDeviceMemory handle

flags must be a valid combination of VkSparseMemoryBindFlagBits values

See Also

VkDeviceMemory, VkDeviceSize, VkExtent3D, VkImageSubresource, VkOffset3D, VkSparseImageMemoryBindInfo, VkSparseMemoryBindFlags

Constructors

SparseImageMemoryBind

Fields

subresource :: ImageSubresource

offset :: Offset3D

extent :: Extent3D

memory :: DeviceMemory

memoryOffset :: DeviceSize

flags :: SparseMemoryBindFlags

withCStructSparseImageMemoryBindInfo :: SparseImageMemoryBindInfo -> (VkSparseImageMemoryBindInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkSparseImageMemoryBindInfo and marshal a SparseImageMemoryBindInfo into it. The VkSparseImageMemoryBindInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructSparseImageMemoryBindInfo :: VkSparseImageMemoryBindInfo -> IO SparseImageMemoryBindInfo

A function to read a VkSparseImageMemoryBindInfo and all additional structures in the pointer chain into a SparseImageMemoryBindInfo.

data SparseImageMemoryBindInfo

VkSparseImageMemoryBindInfo - Structure specifying sparse image memory bind info

Valid Usage

The subresource.mipLevel member of each element of pBinds must be less than the mipLevels specified in VkImageCreateInfo when image was created

The subresource.arrayLayer member of each element of pBinds must be less than the arrayLayers specified in VkImageCreateInfo when image was created

Valid Usage (Implicit)

image must be a valid VkImage handle

pBinds must be a valid pointer to an array of bindCount valid VkSparseImageMemoryBind structures

bindCount must be greater than 0

See Also

VkBindSparseInfo, VkImage, VkSparseImageMemoryBind

Constructors

SparseImageMemoryBindInfo

Fields

image :: Image

binds :: Vector SparseImageMemoryBind

withCStructSparseImageMemoryRequirements :: SparseImageMemoryRequirements -> (VkSparseImageMemoryRequirements -> IO a) -> IO a

A function to temporarily allocate memory for a VkSparseImageMemoryRequirements and marshal a SparseImageMemoryRequirements into it. The VkSparseImageMemoryRequirements is only valid inside the provided computation and must not be returned out of it.

fromCStructSparseImageMemoryRequirements :: VkSparseImageMemoryRequirements -> IO SparseImageMemoryRequirements

A function to read a VkSparseImageMemoryRequirements and all additional structures in the pointer chain into a SparseImageMemoryRequirements.

data SparseImageMemoryRequirements

VkSparseImageMemoryRequirements - Structure specifying sparse image memory requirements

See Also

VkDeviceSize, VkSparseImageFormatProperties, VkSparseImageMemoryRequirements2, vkGetImageSparseMemoryRequirements

Constructors

SparseImageMemoryRequirements

Fields

formatProperties :: SparseImageFormatProperties

imageMipTailFirstLod :: Word32

imageMipTailSize :: DeviceSize

imageMipTailOffset :: DeviceSize

imageMipTailStride :: DeviceSize

withCStructSparseImageOpaqueMemoryBindInfo :: SparseImageOpaqueMemoryBindInfo -> (VkSparseImageOpaqueMemoryBindInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkSparseImageOpaqueMemoryBindInfo and marshal a SparseImageOpaqueMemoryBindInfo into it. The VkSparseImageOpaqueMemoryBindInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructSparseImageOpaqueMemoryBindInfo :: VkSparseImageOpaqueMemoryBindInfo -> IO SparseImageOpaqueMemoryBindInfo

A function to read a VkSparseImageOpaqueMemoryBindInfo and all additional structures in the pointer chain into a SparseImageOpaqueMemoryBindInfo.

data SparseImageOpaqueMemoryBindInfo

VkSparseImageOpaqueMemoryBindInfo - Structure specifying sparse image opaque memory bind info

Valid Usage

If the flags member of any element of pBinds contains VK_SPARSE_MEMORY_BIND_METADATA_BIT, the binding range defined must be within the mip tail region of the metadata aspect of image

Valid Usage (Implicit)

image must be a valid VkImage handle

pBinds must be a valid pointer to an array of bindCount valid VkSparseMemoryBind structures

bindCount must be greater than 0

See Also

VkBindSparseInfo, VkImage, VkSparseMemoryBind

Constructors

SparseImageOpaqueMemoryBindInfo

Fields

image :: Image

binds :: Vector SparseMemoryBind

withCStructSparseMemoryBind :: SparseMemoryBind -> (VkSparseMemoryBind -> IO a) -> IO a

A function to temporarily allocate memory for a VkSparseMemoryBind and marshal a SparseMemoryBind into it. The VkSparseMemoryBind is only valid inside the provided computation and must not be returned out of it.

fromCStructSparseMemoryBind :: VkSparseMemoryBind -> IO SparseMemoryBind

A function to read a VkSparseMemoryBind and all additional structures in the pointer chain into a SparseMemoryBind.

data SparseMemoryBind

VkSparseMemoryBind - Structure specifying a sparse memory bind operation

Description

The binding range [resourceOffset, resourceOffset + size) has different constraints based on flags. If flags contains VK_SPARSE_MEMORY_BIND_METADATA_BIT, the binding range must be within the mip tail region of the metadata aspect. This metadata region is defined by:

metadataRegion = [base, base + imageMipTailSize)

base = imageMipTailOffset + imageMipTailStride × n

and imageMipTailOffset, imageMipTailSize, and imageMipTailStride values are from the VkSparseImageMemoryRequirements corresponding to the metadata aspect of the image, and n is a valid array layer index for the image,

imageMipTailStride is considered to be zero for aspects where VkSparseImageMemoryRequirements::formatProperties.flags contains VK_SPARSE_IMAGE_FORMAT_SINGLE_MIPTAIL_BIT.

If flags does not contain VK_SPARSE_MEMORY_BIND_METADATA_BIT, the binding range must be within the range [0,VkMemoryRequirements::size).

Valid Usage

If memory is not VK_NULL_HANDLE, memory and memoryOffset must match the memory requirements of the resource, as described in section https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#resources-association

If memory is not VK_NULL_HANDLE, memory must not have been created with a memory type that reports VK_MEMORY_PROPERTY_LAZILY_ALLOCATED_BIT bit set

size must be greater than 0

resourceOffset must be less than the size of the resource

size must be less than or equal to the size of the resource minus resourceOffset

memoryOffset must be less than the size of memory

size must be less than or equal to the size of memory minus memoryOffset

Valid Usage (Implicit)

If memory is not VK_NULL_HANDLE, memory must be a valid VkDeviceMemory handle

flags must be a valid combination of VkSparseMemoryBindFlagBits values

See Also

VkDeviceMemory, VkDeviceSize, VkSparseBufferMemoryBindInfo, VkSparseImageOpaqueMemoryBindInfo, VkSparseMemoryBindFlags

Constructors

SparseMemoryBind

Fields

resourceOffset :: DeviceSize

size :: DeviceSize

memory :: DeviceMemory

memoryOffset :: DeviceSize

flags :: SparseMemoryBindFlags

type SparseMemoryBindFlagBits = VkSparseMemoryBindFlagBits

VkSparseMemoryBindFlagBits - Bitmask specifying usage of a sparse memory binding operation

See Also

VkSparseMemoryBindFlags

pattern SPARSE_MEMORY_BIND_METADATA_BIT :: a ~ SparseMemoryBindFlagBits => a

VK_SPARSE_MEMORY_BIND_METADATA_BIT specifies that the memory being bound is only for the metadata aspect.

type SparseMemoryBindFlags = SparseMemoryBindFlagBits

VkSparseMemoryBindFlags - Bitmask of VkSparseMemoryBindFlagBits

Description

VkSparseMemoryBindFlags is a bitmask type for setting a mask of zero or more VkSparseMemoryBindFlagBits.

See Also

VkSparseImageMemoryBind, VkSparseMemoryBind, VkSparseMemoryBindFlagBits

getNumImageSparseMemoryRequirements :: Device -> Image -> IO Word32

vkGetImageSparseMemoryRequirements - Query the memory requirements for a sparse image

Parameters

device is the logical device that owns the image.

image is the VkImage object to get the memory requirements for.

pSparseMemoryRequirementCount is a pointer to an integer related to the number of sparse memory requirements available or queried, as described below.

pSparseMemoryRequirements is either NULL or a pointer to an array of VkSparseImageMemoryRequirements structures.

Description

If pSparseMemoryRequirements is NULL, then the number of sparse memory requirements available is returned in pSparseMemoryRequirementCount. Otherwise, pSparseMemoryRequirementCount must point to a variable set by the user to the number of elements in the pSparseMemoryRequirements array, and on return the variable is overwritten with the number of structures actually written to pSparseMemoryRequirements. If pSparseMemoryRequirementCount is less than the number of sparse memory requirements available, at most pSparseMemoryRequirementCount structures will be written.

If the image was not created with VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT then pSparseMemoryRequirementCount will be set to zero and pSparseMemoryRequirements will not be written to.

Note

It is legal for an implementation to report a larger value in VkMemoryRequirements::size than would be obtained by adding together memory sizes for all VkSparseImageMemoryRequirements returned by vkGetImageSparseMemoryRequirements. This may occur when the implementation requires unused padding in the address range describing the resource.

Valid Usage (Implicit)

device must be a valid VkDevice handle

image must be a valid VkImage handle

pSparseMemoryRequirementCount must be a valid pointer to a uint32_t value

If the value referenced by pSparseMemoryRequirementCount is not 0, and pSparseMemoryRequirements is not NULL, pSparseMemoryRequirements must be a valid pointer to an array of pSparseMemoryRequirementCount VkSparseImageMemoryRequirements structures

image must have been created, allocated, or retrieved from device

See Also

VkDevice, VkImage, VkSparseImageMemoryRequirements

getImageSparseMemoryRequirements :: Device -> Image -> Word32 -> IO (Vector SparseImageMemoryRequirements)

vkGetImageSparseMemoryRequirements - Query the memory requirements for a sparse image

Parameters

device is the logical device that owns the image.

image is the VkImage object to get the memory requirements for.

pSparseMemoryRequirementCount is a pointer to an integer related to the number of sparse memory requirements available or queried, as described below.

pSparseMemoryRequirements is either NULL or a pointer to an array of VkSparseImageMemoryRequirements structures.

Description

If pSparseMemoryRequirements is NULL, then the number of sparse memory requirements available is returned in pSparseMemoryRequirementCount. Otherwise, pSparseMemoryRequirementCount must point to a variable set by the user to the number of elements in the pSparseMemoryRequirements array, and on return the variable is overwritten with the number of structures actually written to pSparseMemoryRequirements. If pSparseMemoryRequirementCount is less than the number of sparse memory requirements available, at most pSparseMemoryRequirementCount structures will be written.

If the image was not created with VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT then pSparseMemoryRequirementCount will be set to zero and pSparseMemoryRequirements will not be written to.

Note

It is legal for an implementation to report a larger value in VkMemoryRequirements::size than would be obtained by adding together memory sizes for all VkSparseImageMemoryRequirements returned by vkGetImageSparseMemoryRequirements. This may occur when the implementation requires unused padding in the address range describing the resource.

Valid Usage (Implicit)

device must be a valid VkDevice handle

image must be a valid VkImage handle

pSparseMemoryRequirementCount must be a valid pointer to a uint32_t value

If the value referenced by pSparseMemoryRequirementCount is not 0, and pSparseMemoryRequirements is not NULL, pSparseMemoryRequirements must be a valid pointer to an array of pSparseMemoryRequirementCount VkSparseImageMemoryRequirements structures

image must have been created, allocated, or retrieved from device

See Also

VkDevice, VkImage, VkSparseImageMemoryRequirements

getAllImageSparseMemoryRequirements :: Device -> Image -> IO (Vector SparseImageMemoryRequirements)

Returns all the values available from getImageSparseMemoryRequirements.

getNumPhysicalDeviceSparseImageFormatProperties :: PhysicalDevice -> Format -> ImageType -> SampleCountFlagBits -> ImageUsageFlags -> ImageTiling -> IO Word32

vkGetPhysicalDeviceSparseImageFormatProperties - Retrieve properties of an image format applied to sparse images

Parameters

physicalDevice is the physical device from which to query the sparse image capabilities.

format is the image format.

type is the dimensionality of image.

samples is the number of samples per texel as defined in VkSampleCountFlagBits.

usage is a bitmask describing the intended usage of the image.

tiling is the tiling arrangement of the texel blocks in memory.

pPropertyCount is a pointer to an integer related to the number of sparse format properties available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkSparseImageFormatProperties structures.

Description

If pProperties is NULL, then the number of sparse format properties available is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If pPropertyCount is less than the number of sparse format properties available, at most pPropertyCount structures will be written.

If VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT is not supported for the given arguments, pPropertyCount will be set to zero upon return, and no data will be written to pProperties.

Multiple aspects are returned for depth/stencil images that are implemented as separate planes by the implementation. The depth and stencil data planes each have unique VkSparseImageFormatProperties data.

Depth/stencil images with depth and stencil data interleaved into a single plane will return a single VkSparseImageFormatProperties structure with the aspectMask set to VK_IMAGE_ASPECT_DEPTH_BIT | VK_IMAGE_ASPECT_STENCIL_BIT.

Valid Usage

samples must be a bit value that is set in VkImageFormatProperties::sampleCounts returned by vkGetPhysicalDeviceImageFormatProperties with format, type, tiling, and usage equal to those in this command and flags equal to the value that is set in VkImageCreateInfo::flags when the image is created

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

format must be a valid VkFormat value

type must be a valid VkImageType value

samples must be a valid VkSampleCountFlagBits value

usage must be a valid combination of VkImageUsageFlagBits values

usage must not be 0

tiling must be a valid VkImageTiling value

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkSparseImageFormatProperties structures

See Also

VkFormat, VkImageTiling, VkImageType, VkImageUsageFlags, VkPhysicalDevice, VkSampleCountFlagBits, VkSparseImageFormatProperties

getPhysicalDeviceSparseImageFormatProperties :: PhysicalDevice -> Format -> ImageType -> SampleCountFlagBits -> ImageUsageFlags -> ImageTiling -> Word32 -> IO (Vector SparseImageFormatProperties)

vkGetPhysicalDeviceSparseImageFormatProperties - Retrieve properties of an image format applied to sparse images

Parameters

physicalDevice is the physical device from which to query the sparse image capabilities.

format is the image format.

type is the dimensionality of image.

samples is the number of samples per texel as defined in VkSampleCountFlagBits.

usage is a bitmask describing the intended usage of the image.

tiling is the tiling arrangement of the texel blocks in memory.

pPropertyCount is a pointer to an integer related to the number of sparse format properties available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkSparseImageFormatProperties structures.

Description

If pProperties is NULL, then the number of sparse format properties available is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If pPropertyCount is less than the number of sparse format properties available, at most pPropertyCount structures will be written.

If VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT is not supported for the given arguments, pPropertyCount will be set to zero upon return, and no data will be written to pProperties.

Multiple aspects are returned for depth/stencil images that are implemented as separate planes by the implementation. The depth and stencil data planes each have unique VkSparseImageFormatProperties data.

Depth/stencil images with depth and stencil data interleaved into a single plane will return a single VkSparseImageFormatProperties structure with the aspectMask set to VK_IMAGE_ASPECT_DEPTH_BIT | VK_IMAGE_ASPECT_STENCIL_BIT.

Valid Usage

samples must be a bit value that is set in VkImageFormatProperties::sampleCounts returned by vkGetPhysicalDeviceImageFormatProperties with format, type, tiling, and usage equal to those in this command and flags equal to the value that is set in VkImageCreateInfo::flags when the image is created

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

format must be a valid VkFormat value

type must be a valid VkImageType value

samples must be a valid VkSampleCountFlagBits value

usage must be a valid combination of VkImageUsageFlagBits values

usage must not be 0

tiling must be a valid VkImageTiling value

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkSparseImageFormatProperties structures

See Also

VkFormat, VkImageTiling, VkImageType, VkImageUsageFlags, VkPhysicalDevice, VkSampleCountFlagBits, VkSparseImageFormatProperties

getAllPhysicalDeviceSparseImageFormatProperties :: PhysicalDevice -> Format -> ImageType -> SampleCountFlagBits -> ImageUsageFlags -> ImageTiling -> IO (Vector SparseImageFormatProperties)

Returns all the values available from getPhysicalDeviceSparseImageFormatProperties.

queueBindSparse :: Queue -> Vector BindSparseInfo -> Fence -> IO ()

vkQueueBindSparse - Bind device memory to a sparse resource object

Parameters

queue is the queue that the sparse binding operations will be submitted to.

bindInfoCount is the number of elements in the pBindInfo array.

pBindInfo is an array of VkBindSparseInfo structures, each specifying a sparse binding submission batch.

fence is an optional handle to a fence to be signaled. If fence is not VK_NULL_HANDLE, it defines a fence signal operation.

Description

vkQueueBindSparse is a queue submission command, with each batch defined by an element of pBindInfo as an instance of the VkBindSparseInfo structure. Batches begin execution in the order they appear in pBindInfo, but may complete out of order.

Within a batch, a given range of a resource must not be bound more than once. Across batches, if a range is to be bound to one allocation and offset and then to another allocation and offset, then the application must guarantee (usually using semaphores) that the binding operations are executed in the correct order, as well as to order binding operations against the execution of command buffer submissions.

As no operation to vkQueueBindSparse causes any pipeline stage to access memory, synchronization primitives used in this command effectively only define execution dependencies.

Additional information about fence and semaphore operation is described in the synchronization chapter.

Valid Usage

If fence is not VK_NULL_HANDLE, fence must be unsignaled

If fence is not VK_NULL_HANDLE, fence must not be associated with any other queue command that has not yet completed execution on that queue

Each element of the pSignalSemaphores member of each element of pBindInfo must be unsignaled when the semaphore signal operation it defines is executed on the device

When a semaphore unsignal operation defined by any element of the pWaitSemaphores member of any element of pBindInfo executes on queue, no other queue must be waiting on the same semaphore.

All elements of the pWaitSemaphores member of all elements of pBindInfo must be semaphores that are signaled, or have semaphore signal operations previously submitted for execution.

Valid Usage (Implicit)

queue must be a valid VkQueue handle

If bindInfoCount is not 0, pBindInfo must be a valid pointer to an array of bindInfoCount valid VkBindSparseInfo structures

If fence is not VK_NULL_HANDLE, fence must be a valid VkFence handle

The queue must support sparse binding operations

Both of fence, and queue that are valid handles must have been created, allocated, or retrieved from the same VkDevice

Host Synchronization

Host access to queue must be externally synchronized

Host access to pBindInfo[].pWaitSemaphores[] must be externally synchronized

Host access to pBindInfo[].pSignalSemaphores[] must be externally synchronized

Host access to pBindInfo[].pBufferBinds[].buffer must be externally synchronized

Host access to pBindInfo[].pImageOpaqueBinds[].image must be externally synchronized

Host access to pBindInfo[].pImageBinds[].image must be externally synchronized

Host access to fence must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| -               | -               | SPARSE_BINDING  | -               |
+-----------------+-----------------+-----------------+-----------------+

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_DEVICE_LOST

See Also

VkBindSparseInfo, VkFence, VkQueue