Graphics.Vulkan.Core11.Promoted_From_VK_KHR_protected_memory

Documentation

withCStructDeviceQueueInfo2 :: DeviceQueueInfo2 -> (VkDeviceQueueInfo2 -> IO a) -> IO a

A function to temporarily allocate memory for a VkDeviceQueueInfo2 and marshal a DeviceQueueInfo2 into it. The VkDeviceQueueInfo2 is only valid inside the provided computation and must not be returned out of it.

fromCStructDeviceQueueInfo2 :: VkDeviceQueueInfo2 -> IO DeviceQueueInfo2

A function to read a VkDeviceQueueInfo2 and all additional structures in the pointer chain into a DeviceQueueInfo2.

data DeviceQueueInfo2

VkDeviceQueueInfo2 - Structure specifying the parameters used for device queue creation

Description

The queue returned by vkGetDeviceQueue2 must have the same flags value from this structure as that used at device creation time in a VkDeviceQueueCreateInfo instance. If no matching flags were specified at device creation time then pQueue will return VK_NULL_HANDLE.

Valid Usage (Implicit)

See Also

VkDeviceQueueCreateFlags, VkStructureType, vkGetDeviceQueue2

Constructors

DeviceQueueInfo2

Fields

next :: Maybe SomeVkStruct

flags :: DeviceQueueCreateFlags

queueFamilyIndex :: Word32

queueIndex :: Word32

withCStructPhysicalDeviceProtectedMemoryFeatures :: PhysicalDeviceProtectedMemoryFeatures -> (VkPhysicalDeviceProtectedMemoryFeatures -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceProtectedMemoryFeatures and marshal a PhysicalDeviceProtectedMemoryFeatures into it. The VkPhysicalDeviceProtectedMemoryFeatures is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceProtectedMemoryFeatures :: VkPhysicalDeviceProtectedMemoryFeatures -> IO PhysicalDeviceProtectedMemoryFeatures

A function to read a VkPhysicalDeviceProtectedMemoryFeatures and all additional structures in the pointer chain into a PhysicalDeviceProtectedMemoryFeatures.

data PhysicalDeviceProtectedMemoryFeatures

VkPhysicalDeviceProtectedMemoryFeatures - Structure describing protected memory features that can be supported by an implementation

Description

If the VkPhysicalDeviceProtectedMemoryFeatures structure is included in the pNext chain of VkPhysicalDeviceFeatures2, it is filled with a value indicating whether the feature is supported.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceProtectedMemoryFeatures

Fields

next :: Maybe SomeVkStruct

protectedMemory :: Bool

withCStructPhysicalDeviceProtectedMemoryProperties :: PhysicalDeviceProtectedMemoryProperties -> (VkPhysicalDeviceProtectedMemoryProperties -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceProtectedMemoryProperties and marshal a PhysicalDeviceProtectedMemoryProperties into it. The VkPhysicalDeviceProtectedMemoryProperties is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceProtectedMemoryProperties :: VkPhysicalDeviceProtectedMemoryProperties -> IO PhysicalDeviceProtectedMemoryProperties

A function to read a VkPhysicalDeviceProtectedMemoryProperties and all additional structures in the pointer chain into a PhysicalDeviceProtectedMemoryProperties.

data PhysicalDeviceProtectedMemoryProperties

VkPhysicalDeviceProtectedMemoryProperties - Structure describing protected memory properties that can be supported by an implementation

Description

If the VkPhysicalDeviceProtectedMemoryProperties structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with a value indicating the implementation-dependent behavior.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceProtectedMemoryProperties

Fields

next :: Maybe SomeVkStruct

protectedNoFault :: Bool

withCStructProtectedSubmitInfo :: ProtectedSubmitInfo -> (VkProtectedSubmitInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkProtectedSubmitInfo and marshal a ProtectedSubmitInfo into it. The VkProtectedSubmitInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructProtectedSubmitInfo :: VkProtectedSubmitInfo -> IO ProtectedSubmitInfo

A function to read a VkProtectedSubmitInfo and all additional structures in the pointer chain into a ProtectedSubmitInfo.

data ProtectedSubmitInfo

VkProtectedSubmitInfo - Structure indicating whether the submission is protected

Valid Usage

If the protected memory feature is not enabled, protectedSubmit must not be VK_TRUE.

If protectedSubmit is VK_TRUE, then each element of the pCommandBuffers array must be a protected command buffer.

If protectedSubmit is VK_FALSE, then each element of the pCommandBuffers array must be an unprotected command buffer.

If the VkSubmitInfo::pNext chain does not include a VkProtectedSubmitInfo structure, then each element of the command buffer of the pCommandBuffers array must be an unprotected command buffer.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PROTECTED_SUBMIT_INFO

See Also

VkBool32, VkStructureType

Constructors

ProtectedSubmitInfo

Fields

next :: Maybe SomeVkStruct

protectedSubmit :: Bool

getDeviceQueue2 :: Device -> DeviceQueueInfo2 -> IO Queue

vkGetDeviceQueue2 - Get a queue handle from a device

Parameters

device is the logical device that owns the queue.

pQueueInfo points to an instance of the VkDeviceQueueInfo2 structure, describing the parameters used to create the device queue.

pQueue is a pointer to a VkQueue object that will be filled with the handle for the requested queue.

Valid Usage (Implicit)

See Also

VkDevice, VkDeviceQueueInfo2, VkQueue

pattern STRUCTURE_TYPE_PROTECTED_SUBMIT_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_PROTECTED_MEMORY_FEATURES :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_PROTECTED_MEMORY_PROPERTIES :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEVICE_QUEUE_INFO_2 :: a ~ StructureType => a

pattern QUEUE_PROTECTED_BIT :: a ~ QueueFlagBits => a

pattern DEVICE_QUEUE_CREATE_PROTECTED_BIT :: a ~ DeviceQueueCreateFlagBits => a

VK_DEVICE_QUEUE_CREATE_PROTECTED_BIT specifies that the device queue is a protected-capable queue. If the protected memory feature is not enabled, the VK_DEVICE_QUEUE_CREATE_PROTECTED_BIT bit of flags must not be set.

pattern MEMORY_PROPERTY_PROTECTED_BIT :: a ~ MemoryPropertyFlagBits => a

pattern BUFFER_CREATE_PROTECTED_BIT :: a ~ BufferCreateFlagBits => a

pattern IMAGE_CREATE_PROTECTED_BIT :: a ~ ImageCreateFlagBits => a

pattern COMMAND_POOL_CREATE_PROTECTED_BIT :: a ~ CommandPoolCreateFlagBits => a