Graphics.Vulkan.Core11.Promoted_From_VK_KHR_subgroup

Documentation

withCStructPhysicalDeviceSubgroupProperties :: PhysicalDeviceSubgroupProperties -> (VkPhysicalDeviceSubgroupProperties -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceSubgroupProperties and marshal a PhysicalDeviceSubgroupProperties into it. The VkPhysicalDeviceSubgroupProperties is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceSubgroupProperties :: VkPhysicalDeviceSubgroupProperties -> IO PhysicalDeviceSubgroupProperties

A function to read a VkPhysicalDeviceSubgroupProperties and all additional structures in the pointer chain into a PhysicalDeviceSubgroupProperties.

data PhysicalDeviceSubgroupProperties

VkPhysicalDeviceSubgroupProperties - Structure describing subgroup support for an implementation

Members

The members of the VkPhysicalDeviceSubgroupProperties structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceSubgroupProperties structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkBool32, VkShaderStageFlags, VkStructureType, VkSubgroupFeatureFlags

Constructors

PhysicalDeviceSubgroupProperties

Fields

next :: Maybe SomeVkStruct

subgroupSize :: Word32

supportedStages :: ShaderStageFlags

supportedOperations :: SubgroupFeatureFlags

quadOperationsInAllStages :: Bool

type SubgroupFeatureFlagBits = VkSubgroupFeatureFlagBits

VkSubgroupFeatureFlagBits - Enum describing what subgroup operations are supported

See Also

VkSubgroupFeatureFlags

pattern SUBGROUP_FEATURE_BASIC_BIT :: a ~ SubgroupFeatureFlagBits => a

VK_SUBGROUP_FEATURE_BASIC_BIT specifies the device will accept SPIR-V shader modules that contain the GroupNonUniform capability.

pattern SUBGROUP_FEATURE_VOTE_BIT :: a ~ SubgroupFeatureFlagBits => a

VK_SUBGROUP_FEATURE_VOTE_BIT specifies the device will accept SPIR-V shader modules that contain the GroupNonUniformVote capability.

pattern SUBGROUP_FEATURE_ARITHMETIC_BIT :: a ~ SubgroupFeatureFlagBits => a

VK_SUBGROUP_FEATURE_ARITHMETIC_BIT specifies the device will accept SPIR-V shader modules that contain the GroupNonUniformArithmetic capability.

pattern SUBGROUP_FEATURE_BALLOT_BIT :: a ~ SubgroupFeatureFlagBits => a

VK_SUBGROUP_FEATURE_BALLOT_BIT specifies the device will accept SPIR-V shader modules that contain the GroupNonUniformBallot capability.

pattern SUBGROUP_FEATURE_SHUFFLE_BIT :: a ~ SubgroupFeatureFlagBits => a

VK_SUBGROUP_FEATURE_SHUFFLE_BIT specifies the device will accept SPIR-V shader modules that contain the GroupNonUniformShuffle capability.

pattern SUBGROUP_FEATURE_SHUFFLE_RELATIVE_BIT :: a ~ SubgroupFeatureFlagBits => a

VK_SUBGROUP_FEATURE_SHUFFLE_RELATIVE_BIT specifies the device will accept SPIR-V shader modules that contain the GroupNonUniformShuffleRelative capability.

pattern SUBGROUP_FEATURE_CLUSTERED_BIT :: a ~ SubgroupFeatureFlagBits => a

VK_SUBGROUP_FEATURE_CLUSTERED_BIT specifies the device will accept SPIR-V shader modules that contain the GroupNonUniformClustered capability.

pattern SUBGROUP_FEATURE_QUAD_BIT :: a ~ SubgroupFeatureFlagBits => a

VK_SUBGROUP_FEATURE_QUAD_BIT specifies the device will accept SPIR-V shader modules that contain the GroupNonUniformQuad capability.

pattern SUBGROUP_FEATURE_PARTITIONED_BIT_NV :: a ~ SubgroupFeatureFlagBits => a

type SubgroupFeatureFlags = SubgroupFeatureFlagBits

VkSubgroupFeatureFlags - Bitmask of VkSubgroupFeatureFlagBits

Description

VkSubgroupFeatureFlags is a bitmask type for setting a mask of zero or more VkSubgroupFeatureFlagBits.

See Also

VkPhysicalDeviceSubgroupProperties, VkSubgroupFeatureFlagBits

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SUBGROUP_PROPERTIES :: a ~ StructureType => a