Graphics.Vulkan.Core11.Promoted_from_VK_KHR_dedicated_allocation

Documentation

withCStructMemoryDedicatedAllocateInfo :: MemoryDedicatedAllocateInfo -> (VkMemoryDedicatedAllocateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkMemoryDedicatedAllocateInfo and marshal a MemoryDedicatedAllocateInfo into it. The VkMemoryDedicatedAllocateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructMemoryDedicatedAllocateInfo :: VkMemoryDedicatedAllocateInfo -> IO MemoryDedicatedAllocateInfo

A function to read a VkMemoryDedicatedAllocateInfo and all additional structures in the pointer chain into a MemoryDedicatedAllocateInfo.

data MemoryDedicatedAllocateInfo

VkMemoryDedicatedAllocateInfo - Specify a dedicated memory allocation resource

Valid Usage

At least one of image and buffer must be VK_NULL_HANDLE

If image is not VK_NULL_HANDLE, VkMemoryAllocateInfo::allocationSize must equal the VkMemoryRequirements::size of the image

If image is not VK_NULL_HANDLE, image must have been created without VK_IMAGE_CREATE_SPARSE_BINDING_BIT set in VkImageCreateInfo::flags

If buffer is not VK_NULL_HANDLE, VkMemoryAllocateInfo::allocationSize must equal the VkMemoryRequirements::size of the buffer

If buffer is not VK_NULL_HANDLE, buffer must have been created without VK_BUFFER_CREATE_SPARSE_BINDING_BIT set in VkBufferCreateInfo::flags

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_MEMORY_DEDICATED_ALLOCATE_INFO

If image is not VK_NULL_HANDLE, image must be a valid VkImage handle

If buffer is not VK_NULL_HANDLE, buffer must be a valid VkBuffer handle

Both of buffer, and image that are valid handles must have been created, allocated, or retrieved from the same VkDevice

See Also

VkBuffer, VkImage, VkStructureType

Constructors

MemoryDedicatedAllocateInfo

Fields

next :: Maybe SomeVkStruct

image :: Image

buffer :: Buffer

withCStructMemoryDedicatedRequirements :: MemoryDedicatedRequirements -> (VkMemoryDedicatedRequirements -> IO a) -> IO a

A function to temporarily allocate memory for a VkMemoryDedicatedRequirements and marshal a MemoryDedicatedRequirements into it. The VkMemoryDedicatedRequirements is only valid inside the provided computation and must not be returned out of it.

fromCStructMemoryDedicatedRequirements :: VkMemoryDedicatedRequirements -> IO MemoryDedicatedRequirements

A function to read a VkMemoryDedicatedRequirements and all additional structures in the pointer chain into a MemoryDedicatedRequirements.

data MemoryDedicatedRequirements

VkMemoryDedicatedRequirements - Structure describing dedicated allocation requirements of buffer and image resources

Description

When the implementation sets requiresDedicatedAllocation to VK_TRUE, it must also set prefersDedicatedAllocation to VK_TRUE.

If the VkMemoryDedicatedRequirements structure is included in the pNext chain of the VkMemoryRequirements2 structure passed as the pMemoryRequirements parameter of a vkGetBufferMemoryRequirements2 call, requiresDedicatedAllocation may be VK_TRUE under one of the following conditions:

none

In all other cases, requiresDedicatedAllocation must be set to VK_FALSE by the implementation whenever a VkMemoryDedicatedRequirements structure is included in the pNext chain of the VkMemoryRequirements2 structure passed to a call to vkGetBufferMemoryRequirements2.

If the VkMemoryDedicatedRequirements structure is included in the pNext chain of the VkMemoryRequirements2 structure passed as the pMemoryRequirements parameter of a vkGetBufferMemoryRequirements2 call and VK_BUFFER_CREATE_SPARSE_BINDING_BIT was set in VkBufferCreateInfo::flags when buffer was created then the implementation must set both prefersDedicatedAllocation and requiresDedicatedAllocation to VK_FALSE.

If the VkMemoryDedicatedRequirements structure is included in the pNext chain of the VkMemoryRequirements2 structure passed as the pMemoryRequirements parameter of a vkGetImageMemoryRequirements2 call, requiresDedicatedAllocation may be VK_TRUE under one of the following conditions:

none

In all other cases, requiresDedicatedAllocation must be set to VK_FALSE by the implementation whenever a VkMemoryDedicatedRequirements structure is included in the pNext chain of the VkMemoryRequirements2 structure passed to a call to vkGetImageMemoryRequirements2.

If the VkMemoryDedicatedRequirements structure is included in the pNext chain of the VkMemoryRequirements2 structure passed as the pMemoryRequirements parameter of a vkGetImageMemoryRequirements2 call and VK_IMAGE_CREATE_SPARSE_BINDING_BIT was set in VkImageCreateInfo::flags when image was created then the implementation must set both prefersDedicatedAllocation and requiresDedicatedAllocation to VK_FALSE.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_MEMORY_DEDICATED_REQUIREMENTS

See Also

VkBool32, VkStructureType

Constructors

MemoryDedicatedRequirements

Fields

next :: Maybe SomeVkStruct

prefersDedicatedAllocation :: Bool

requiresDedicatedAllocation :: Bool

pattern STRUCTURE_TYPE_MEMORY_DEDICATED_REQUIREMENTS :: a ~ StructureType => a

pattern STRUCTURE_TYPE_MEMORY_DEDICATED_ALLOCATE_INFO :: a ~ StructureType => a