Graphics.Vulkan.Core11.Promoted_from_VK_KHR_descriptor_update_template

Documentation

type DescriptorUpdateTemplate = VkDescriptorUpdateTemplate

VkDescriptorUpdateTemplate - Opaque handle to a descriptor update template

See Also

vkCmdPushDescriptorSetWithTemplateKHR, vkCreateDescriptorUpdateTemplate, vkCreateDescriptorUpdateTemplateKHR, vkDestroyDescriptorUpdateTemplate, vkDestroyDescriptorUpdateTemplateKHR, vkUpdateDescriptorSetWithTemplate, vkUpdateDescriptorSetWithTemplateKHR

type DescriptorUpdateTemplateCreateFlags = VkDescriptorUpdateTemplateCreateFlags

VkDescriptorUpdateTemplateCreateFlags - Reserved for future use

Description

VkDescriptorUpdateTemplateCreateFlags is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkDescriptorUpdateTemplateCreateInfo

type DescriptorUpdateTemplateCreateFlagsKHR = DescriptorUpdateTemplateCreateFlags

withCStructDescriptorUpdateTemplateCreateInfo :: DescriptorUpdateTemplateCreateInfo -> (VkDescriptorUpdateTemplateCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkDescriptorUpdateTemplateCreateInfo and marshal a DescriptorUpdateTemplateCreateInfo into it. The VkDescriptorUpdateTemplateCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructDescriptorUpdateTemplateCreateInfo :: VkDescriptorUpdateTemplateCreateInfo -> IO DescriptorUpdateTemplateCreateInfo

A function to read a VkDescriptorUpdateTemplateCreateInfo and all additional structures in the pointer chain into a DescriptorUpdateTemplateCreateInfo.

data DescriptorUpdateTemplateCreateInfo

VkDescriptorUpdateTemplateCreateInfo - Structure specifying parameters of a newly created descriptor update template

Valid Usage

If templateType is VK_DESCRIPTOR_UPDATE_TEMPLATE_TYPE_DESCRIPTOR_SET, descriptorSetLayout must be a valid VkDescriptorSetLayout handle

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_CREATE_INFO

pNext must be NULL

flags must be 0

pDescriptorUpdateEntries must be a valid pointer to an array of descriptorUpdateEntryCount valid VkDescriptorUpdateTemplateEntry structures

templateType must be a valid VkDescriptorUpdateTemplateType value

If descriptorSetLayout is not VK_NULL_HANDLE, descriptorSetLayout must be a valid VkDescriptorSetLayout handle

descriptorUpdateEntryCount must be greater than 0

Both of descriptorSetLayout, and pipelineLayout that are valid handles must have been created, allocated, or retrieved from the same VkDevice

See Also

VkDescriptorSetLayout, VkDescriptorUpdateTemplateCreateFlags, VkDescriptorUpdateTemplateEntry, VkDescriptorUpdateTemplateType, VkPipelineBindPoint, VkPipelineLayout, VkStructureType, vkCreateDescriptorUpdateTemplate, vkCreateDescriptorUpdateTemplateKHR

Constructors

DescriptorUpdateTemplateCreateInfo

Fields

next :: Maybe SomeVkStruct

flags :: DescriptorUpdateTemplateCreateFlags

descriptorUpdateEntries :: Vector DescriptorUpdateTemplateEntry

templateType :: DescriptorUpdateTemplateType

descriptorSetLayout :: DescriptorSetLayout

pipelineBindPoint :: PipelineBindPoint

pipelineLayout :: PipelineLayout

set :: Word32

withCStructDescriptorUpdateTemplateEntry :: DescriptorUpdateTemplateEntry -> (VkDescriptorUpdateTemplateEntry -> IO a) -> IO a

A function to temporarily allocate memory for a VkDescriptorUpdateTemplateEntry and marshal a DescriptorUpdateTemplateEntry into it. The VkDescriptorUpdateTemplateEntry is only valid inside the provided computation and must not be returned out of it.

fromCStructDescriptorUpdateTemplateEntry :: VkDescriptorUpdateTemplateEntry -> IO DescriptorUpdateTemplateEntry

A function to read a VkDescriptorUpdateTemplateEntry and all additional structures in the pointer chain into a DescriptorUpdateTemplateEntry.

data DescriptorUpdateTemplateEntry

VkDescriptorUpdateTemplateEntry - Describes a single descriptor update of the descriptor update template

Valid Usage (Implicit)

See Also

VkDescriptorType, VkDescriptorUpdateTemplateCreateInfo

Constructors

DescriptorUpdateTemplateEntry

Fields

dstBinding :: Word32

dstArrayElement :: Word32

descriptorCount :: Word32

descriptorType :: DescriptorType

offset :: CSize

stride :: CSize

type DescriptorUpdateTemplateType = VkDescriptorUpdateTemplateType

VkDescriptorUpdateTemplateType - Indicates the valid usage of the descriptor update template

See Also

VkDescriptorUpdateTemplateCreateInfo

pattern DESCRIPTOR_UPDATE_TEMPLATE_TYPE_DESCRIPTOR_SET :: a ~ DescriptorUpdateTemplateType => a

VK_DESCRIPTOR_UPDATE_TEMPLATE_TYPE_DESCRIPTOR_SET specifies that the descriptor update template will be used for descriptor set updates only.

pattern DESCRIPTOR_UPDATE_TEMPLATE_TYPE_PUSH_DESCRIPTORS_KHR :: a ~ DescriptorUpdateTemplateType => a

type DescriptorUpdateTemplateTypeKHR = DescriptorUpdateTemplateType

createDescriptorUpdateTemplate :: Device -> DescriptorUpdateTemplateCreateInfo -> Maybe AllocationCallbacks -> IO DescriptorUpdateTemplate

vkCreateDescriptorUpdateTemplate - Create a new descriptor update template

Parameters

device is the logical device that creates the descriptor update template.

pCreateInfo is a pointer to an instance of the VkDescriptorUpdateTemplateCreateInfo structure specifying the set of descriptors to update with a single call to vkUpdateDescriptorSetWithTemplate.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pDescriptorUpdateTemplate points to a VkDescriptorUpdateTemplate handle in which the resulting descriptor update template object is returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkDescriptorUpdateTemplateCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pDescriptorUpdateTemplate must be a valid pointer to a VkDescriptorUpdateTemplate handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDescriptorUpdateTemplate, VkDescriptorUpdateTemplateCreateInfo, VkDevice

destroyDescriptorUpdateTemplate :: Device -> DescriptorUpdateTemplate -> Maybe AllocationCallbacks -> IO ()

vkDestroyDescriptorUpdateTemplate - Destroy a descriptor update template object

Parameters

device is the logical device that has been used to create the descriptor update template

descriptorUpdateTemplate is the descriptor update template to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

If VkAllocationCallbacks were provided when descriptorSetLayout was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when descriptorSetLayout was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid VkDevice handle

If descriptorUpdateTemplate is not VK_NULL_HANDLE, descriptorUpdateTemplate must be a valid VkDescriptorUpdateTemplate handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

If descriptorUpdateTemplate is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to descriptorUpdateTemplate must be externally synchronized

See Also

VkAllocationCallbacks, VkDescriptorUpdateTemplate, VkDevice

updateDescriptorSetWithTemplate :: Device -> DescriptorSet -> DescriptorUpdateTemplate -> Ptr () -> IO ()

vkUpdateDescriptorSetWithTemplate - Update the contents of a descriptor set object using an update template

Parameters

device is the logical device that updates the descriptor sets.

descriptorSet is the descriptor set to update

descriptorUpdateTemplate is the VkDescriptorUpdateTemplate which specifies the update mapping between pData and the descriptor set to update.

pData is a pointer to memory which contains one or more structures of VkDescriptorImageInfo, VkDescriptorBufferInfo, or VkBufferView used to write the descriptors.

Valid Usage

pData must be a valid pointer to a memory that contains one or more valid instances of VkDescriptorImageInfo, VkDescriptorBufferInfo, or VkBufferView in a layout defined by descriptorUpdateTemplate when it was created with vkCreateDescriptorUpdateTemplate

Valid Usage (Implicit)

device must be a valid VkDevice handle

descriptorSet must be a valid VkDescriptorSet handle

descriptorUpdateTemplate must be a valid VkDescriptorUpdateTemplate handle

descriptorUpdateTemplate must have been created, allocated, or retrieved from device

Host Synchronization

Host access to descriptorSet must be externally synchronized

API example.

struct AppBufferView {
    VkBufferView bufferView;
    uint32_t     applicationRelatedInformation;
};

struct AppDataStructure
{
    VkDescriptorImageInfo  imageInfo;          // a single image info
    VkDescriptorBufferInfo bufferInfoArray[3]; // 3 buffer infos in an array
    AppBufferView          bufferView[2];      // An application defined structure containing a bufferView
    // ... some more application related data
};

const VkDescriptorUpdateTemplateEntry descriptorUpdateTemplateEntries[] =
{
    // binding to a single image descriptor
    {
        0,                                           // binding
        0,                                           // dstArrayElement
        1,                                           // descriptorCount
        VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER,   // descriptorType
        offsetof(AppDataStructure, imageInfo),       // offset
        0                                            // stride is not required if descriptorCount is 1
    },

    // binding to an array of buffer descriptors
    {
        1,                                           // binding
        0,                                           // dstArrayElement
        3,                                           // descriptorCount
        VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER,           // descriptorType
        offsetof(AppDataStructure, bufferInfoArray), // offset
        sizeof(VkDescriptorBufferInfo)               // stride, descriptor buffer infos are compact
    },

    // binding to an array of buffer views
    {
        2,                                           // binding
        0,                                           // dstArrayElement
        2,                                           // descriptorCount
        VK_DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER,     // descriptorType
        offsetof(AppDataStructure, bufferView) +
          offsetof(AppBufferView, bufferView),       // offset
        sizeof(AppBufferView)                        // stride, bufferViews do not have to be compact
    },
};

// create a descriptor update template for descriptor set updates
const VkDescriptorUpdateTemplateCreateInfo createInfo =
{
    VK_STRUCTURE_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_CREATE_INFO,  // sType
    NULL,                                                      // pNext
    0,                                                         // flags
    3,                                                         // descriptorUpdateEntryCount
    descriptorUpdateTemplateEntries,                           // pDescriptorUpdateEntries
    VK_DESCRIPTOR_UPDATE_TEMPLATE_TYPE_DESCRIPTOR_SET,         // templateType
    myLayout,                                                  // descriptorSetLayout
    0,                                                         // pipelineBindPoint, ignored by given templateType
    0,                                                         // pipelineLayout, ignored by given templateType
    0,                                                         // set, ignored by given templateType
};

VkDescriptorUpdateTemplate myDescriptorUpdateTemplate;
myResult = vkCreateDescriptorUpdateTemplate(
    myDevice,
    &createInfo,
    NULL,
    &myDescriptorUpdateTemplate);
}

AppDataStructure appData;

// fill appData here or cache it in your engine
vkUpdateDescriptorSetWithTemplate(myDevice, myDescriptorSet, myDescriptorUpdateTemplate, &appData);

See Also

VkDescriptorSet, VkDescriptorUpdateTemplate, VkDevice

withDescriptorUpdateTemplate :: Device -> DescriptorUpdateTemplateCreateInfo -> Maybe AllocationCallbacks -> (DescriptorUpdateTemplate -> IO a) -> IO a

A safe wrapper for createDescriptorUpdateTemplate and destroyDescriptorUpdateTemplate using bracket

The allocated value must not be returned from the provided computation

pattern STRUCTURE_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_CREATE_INFO :: a ~ StructureType => a

pattern OBJECT_TYPE_DESCRIPTOR_UPDATE_TEMPLATE :: a ~ ObjectType => a