Graphics.Vulkan.Core11.Promoted_from_VK_KHR_device_group

Documentation

withCStructDeviceGroupBindSparseInfo :: DeviceGroupBindSparseInfo -> (VkDeviceGroupBindSparseInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkDeviceGroupBindSparseInfo and marshal a DeviceGroupBindSparseInfo into it. The VkDeviceGroupBindSparseInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructDeviceGroupBindSparseInfo :: VkDeviceGroupBindSparseInfo -> IO DeviceGroupBindSparseInfo

A function to read a VkDeviceGroupBindSparseInfo and all additional structures in the pointer chain into a DeviceGroupBindSparseInfo.

data DeviceGroupBindSparseInfo

VkDeviceGroupBindSparseInfo - Structure indicating which instances are bound

Description

These device indices apply to all buffer and image memory binds included in the batch that points to this structure. The semaphore waits and signals for the batch are executed only by the physical device specified by the resourceDeviceIndex.

If this structure is not present, resourceDeviceIndex and memoryDeviceIndex are assumed to be zero.

Valid Usage

resourceDeviceIndex and memoryDeviceIndex must both be valid device indices.

Each memory allocation bound in this batch must have allocated an instance for memoryDeviceIndex.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DEVICE_GROUP_BIND_SPARSE_INFO

See Also

VkStructureType

Constructors

DeviceGroupBindSparseInfo

Fields

next :: Maybe SomeVkStruct

resourceDeviceIndex :: Word32

memoryDeviceIndex :: Word32

withCStructDeviceGroupCommandBufferBeginInfo :: DeviceGroupCommandBufferBeginInfo -> (VkDeviceGroupCommandBufferBeginInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkDeviceGroupCommandBufferBeginInfo and marshal a DeviceGroupCommandBufferBeginInfo into it. The VkDeviceGroupCommandBufferBeginInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructDeviceGroupCommandBufferBeginInfo :: VkDeviceGroupCommandBufferBeginInfo -> IO DeviceGroupCommandBufferBeginInfo

A function to read a VkDeviceGroupCommandBufferBeginInfo and all additional structures in the pointer chain into a DeviceGroupCommandBufferBeginInfo.

data DeviceGroupCommandBufferBeginInfo

VkDeviceGroupCommandBufferBeginInfo - Set the initial device mask for a command buffer

Description

The initial device mask also acts as an upper bound on the set of devices that can ever be in the device mask in the command buffer.

If this structure is not present, the initial value of a command buffer’s device mask is set to include all physical devices in the logical device when the command buffer begins recording.

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

DeviceGroupCommandBufferBeginInfo

Fields

next :: Maybe SomeVkStruct

deviceMask :: Word32

withCStructDeviceGroupRenderPassBeginInfo :: DeviceGroupRenderPassBeginInfo -> (VkDeviceGroupRenderPassBeginInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkDeviceGroupRenderPassBeginInfo and marshal a DeviceGroupRenderPassBeginInfo into it. The VkDeviceGroupRenderPassBeginInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructDeviceGroupRenderPassBeginInfo :: VkDeviceGroupRenderPassBeginInfo -> IO DeviceGroupRenderPassBeginInfo

A function to read a VkDeviceGroupRenderPassBeginInfo and all additional structures in the pointer chain into a DeviceGroupRenderPassBeginInfo.

data DeviceGroupRenderPassBeginInfo

VkDeviceGroupRenderPassBeginInfo - Set the initial device mask and render areas for a render pass instance

Description

The deviceMask serves several purposes. It is an upper bound on the set of physical devices that can be used during the render pass instance, and the initial device mask when the render pass instance begins. In addition, commands transitioning to the next subpass in the render pass instance and commands ending the render pass instance, and, accordingly render pass attachment load, store, and resolve operations and subpass dependencies corresponding to the render pass instance, are executed on the physical devices included in the device mask provided here.

If deviceRenderAreaCount is not zero, then the elements of pDeviceRenderAreas override the value of VkRenderPassBeginInfo::renderArea, and provide a render area specific to each physical device. These render areas serve the same purpose as VkRenderPassBeginInfo::renderArea, including controlling the region of attachments that are cleared by VK_ATTACHMENT_LOAD_OP_CLEAR and that are resolved into resolve attachments.

If this structure is not present, the render pass instance’s device mask is the value of VkDeviceGroupCommandBufferBeginInfo::deviceMask. If this structure is not present or if deviceRenderAreaCount is zero, VkRenderPassBeginInfo::renderArea is used for all physical devices.

Valid Usage

deviceMask must be a valid device mask value

deviceMask must not be zero

deviceMask must be a subset of the command buffer’s initial device mask

deviceRenderAreaCount must either be zero or equal to the number of physical devices in the logical device.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DEVICE_GROUP_RENDER_PASS_BEGIN_INFO

If deviceRenderAreaCount is not 0, pDeviceRenderAreas must be a valid pointer to an array of deviceRenderAreaCount VkRect2D structures

See Also

VkRect2D, VkStructureType

Constructors

DeviceGroupRenderPassBeginInfo

Fields

next :: Maybe SomeVkStruct

deviceMask :: Word32

deviceRenderAreas :: Vector Rect2D

withCStructDeviceGroupSubmitInfo :: DeviceGroupSubmitInfo -> (VkDeviceGroupSubmitInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkDeviceGroupSubmitInfo and marshal a DeviceGroupSubmitInfo into it. The VkDeviceGroupSubmitInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructDeviceGroupSubmitInfo :: VkDeviceGroupSubmitInfo -> IO DeviceGroupSubmitInfo

A function to read a VkDeviceGroupSubmitInfo and all additional structures in the pointer chain into a DeviceGroupSubmitInfo.

data DeviceGroupSubmitInfo

VkDeviceGroupSubmitInfo - Structure indicating which physical devices execute semaphore operations and command buffers

Description

If this structure is not present, semaphore operations and command buffers execute on device index zero.

Valid Usage

waitSemaphoreCount must equal VkSubmitInfo::waitSemaphoreCount

commandBufferCount must equal VkSubmitInfo::commandBufferCount

signalSemaphoreCount must equal VkSubmitInfo::signalSemaphoreCount

All elements of pWaitSemaphoreDeviceIndices and pSignalSemaphoreDeviceIndices must be valid device indices

All elements of pCommandBufferDeviceMasks must be valid device masks

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DEVICE_GROUP_SUBMIT_INFO

If waitSemaphoreCount is not 0, pWaitSemaphoreDeviceIndices must be a valid pointer to an array of waitSemaphoreCount uint32_t values

If commandBufferCount is not 0, pCommandBufferDeviceMasks must be a valid pointer to an array of commandBufferCount uint32_t values

If signalSemaphoreCount is not 0, pSignalSemaphoreDeviceIndices must be a valid pointer to an array of signalSemaphoreCount uint32_t values

See Also

VkStructureType

Constructors

DeviceGroupSubmitInfo

Fields

next :: Maybe SomeVkStruct

waitSemaphoreDeviceIndices :: Vector Word32

commandBufferDeviceMasks :: Vector Word32

signalSemaphoreDeviceIndices :: Vector Word32

type MemoryAllocateFlagBits = VkMemoryAllocateFlagBits

VkMemoryAllocateFlagBits - Bitmask specifying flags for a device memory allocation

See Also

VkMemoryAllocateFlags

pattern MEMORY_ALLOCATE_DEVICE_MASK_BIT :: a ~ MemoryAllocateFlagBits => a

VK_MEMORY_ALLOCATE_DEVICE_MASK_BIT specifies that memory will be allocated for the devices in VkMemoryAllocateFlagsInfo::deviceMask.

type MemoryAllocateFlagBitsKHR = MemoryAllocateFlagBits

type MemoryAllocateFlags = MemoryAllocateFlagBits

VkMemoryAllocateFlags - Bitmask of VkMemoryAllocateFlagBits

Description

VkMemoryAllocateFlags is a bitmask type for setting a mask of zero or more VkMemoryAllocateFlagBits.

See Also

VkMemoryAllocateFlagBits, VkMemoryAllocateFlagsInfo

withCStructMemoryAllocateFlagsInfo :: MemoryAllocateFlagsInfo -> (VkMemoryAllocateFlagsInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkMemoryAllocateFlagsInfo and marshal a MemoryAllocateFlagsInfo into it. The VkMemoryAllocateFlagsInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructMemoryAllocateFlagsInfo :: VkMemoryAllocateFlagsInfo -> IO MemoryAllocateFlagsInfo

A function to read a VkMemoryAllocateFlagsInfo and all additional structures in the pointer chain into a MemoryAllocateFlagsInfo.

data MemoryAllocateFlagsInfo

VkMemoryAllocateFlagsInfo - Structure controlling how many instances of memory will be allocated

Description

If VK_MEMORY_ALLOCATE_DEVICE_MASK_BIT is not set, the number of instances allocated depends on whether VK_MEMORY_HEAP_MULTI_INSTANCE_BIT is set in the memory heap. If VK_MEMORY_HEAP_MULTI_INSTANCE_BIT is set, then memory is allocated for every physical device in the logical device (as if deviceMask has bits set for all device indices). If VK_MEMORY_HEAP_MULTI_INSTANCE_BIT is not set, then a single instance of memory is allocated (as if deviceMask is set to one).

On some implementations, allocations from a multi-instance heap may consume memory on all physical devices even if the deviceMask excludes some devices. If VkPhysicalDeviceGroupProperties::subsetAllocation is VK_TRUE, then memory is only consumed for the devices in the device mask.

Note

In practice, most allocations on a multi-instance heap will be allocated across all physical devices. Unicast allocation support is an optional optimization for a minority of allocations.

Valid Usage

If VK_MEMORY_ALLOCATE_DEVICE_MASK_BIT is set, deviceMask must be a valid device mask.

If VK_MEMORY_ALLOCATE_DEVICE_MASK_BIT is set, deviceMask must not be zero

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_FLAGS_INFO

flags must be a valid combination of VkMemoryAllocateFlagBits values

See Also

VkMemoryAllocateFlags, VkStructureType

Constructors

MemoryAllocateFlagsInfo

Fields

next :: Maybe SomeVkStruct

flags :: MemoryAllocateFlags

deviceMask :: Word32

type MemoryAllocateFlagsKHR = MemoryAllocateFlags

type PeerMemoryFeatureFlagBits = VkPeerMemoryFeatureFlagBits

VkPeerMemoryFeatureFlagBits - Bitmask specifying supported peer memory features

Description

Note

The peer memory features of a memory heap also apply to any accesses that may be performed during image layout transitions.

VK_PEER_MEMORY_FEATURE_COPY_DST_BIT must be supported for all host local heaps and for at least one device local heap.

If a device does not support a peer memory feature, it is still valid to use a resource that includes both local and peer memory bindings with the corresponding access type as long as only the local bindings are actually accessed. For example, an application doing split-frame rendering would use framebuffer attachments that include both local and peer memory bindings, but would scissor the rendering to only update local memory.

See Also

VkPeerMemoryFeatureFlags

pattern PEER_MEMORY_FEATURE_COPY_SRC_BIT :: a ~ PeerMemoryFeatureFlagBits => a

VK_PEER_MEMORY_FEATURE_COPY_SRC_BIT specifies that the memory can be accessed as the source of a vkCmdCopyBuffer, vkCmdCopyImage, vkCmdCopyBufferToImage, or vkCmdCopyImageToBuffer command.

pattern PEER_MEMORY_FEATURE_COPY_DST_BIT :: a ~ PeerMemoryFeatureFlagBits => a

VK_PEER_MEMORY_FEATURE_COPY_DST_BIT specifies that the memory can be accessed as the destination of a vkCmdCopyBuffer, vkCmdCopyImage, vkCmdCopyBufferToImage, or vkCmdCopyImageToBuffer command.

pattern PEER_MEMORY_FEATURE_GENERIC_SRC_BIT :: a ~ PeerMemoryFeatureFlagBits => a

VK_PEER_MEMORY_FEATURE_GENERIC_SRC_BIT specifies that the memory can be read as any memory access type.

pattern PEER_MEMORY_FEATURE_GENERIC_DST_BIT :: a ~ PeerMemoryFeatureFlagBits => a

VK_PEER_MEMORY_FEATURE_GENERIC_DST_BIT specifies that the memory can be written as any memory access type. Shader atomics are considered to be writes.

type PeerMemoryFeatureFlagBitsKHR = PeerMemoryFeatureFlagBits

type PeerMemoryFeatureFlags = PeerMemoryFeatureFlagBits

VkPeerMemoryFeatureFlags - Bitmask of VkPeerMemoryFeatureFlagBits

Description

VkPeerMemoryFeatureFlags is a bitmask type for setting a mask of zero or more VkPeerMemoryFeatureFlagBits.

See Also

VkPeerMemoryFeatureFlagBits, vkGetDeviceGroupPeerMemoryFeatures, vkGetDeviceGroupPeerMemoryFeaturesKHR

type PeerMemoryFeatureFlagsKHR = PeerMemoryFeatureFlags

cmdDispatchBase :: CommandBuffer -> Word32 -> Word32 -> Word32 -> Word32 -> Word32 -> Word32 -> IO ()

vkCmdDispatchBase - Dispatch compute work items

Parameters

commandBuffer is the command buffer into which the command will be recorded.

baseGroupX is the start value for the X component of WorkgroupId.

baseGroupY is the start value for the Y component of WorkgroupId.

baseGroupZ is the start value for the Z component of WorkgroupId.

groupCountX is the number of local workgroups to dispatch in the X dimension.

groupCountY is the number of local workgroups to dispatch in the Y dimension.

groupCountZ is the number of local workgroups to dispatch in the Z dimension.

Description

When the command is executed, a global workgroup consisting of groupCountX × groupCountY × groupCountZ local workgroups is assembled, with WorkgroupId values ranging from [baseGroup*, baseGroup* + groupCount*) in each component. vkCmdDispatch is equivalent to vkCmdDispatchBase(0,0,0,groupCountX,groupCountY,groupCountZ).

Valid Usage

If a VkImageView is sampled with VK_FILTER_LINEAR as a result of this command, then the image view’s format features must contain VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_LINEAR_BIT

If a VkImageView is accessed using atomic operations as a result of this command, then the image view’s format features must contain VK_FORMAT_FEATURE_STORAGE_IMAGE_ATOMIC_BIT

For each set n that is statically used by the VkPipeline bound to the pipeline bind point used by this command, a descriptor set must have been bound to n at the same pipeline bind point, with a VkPipelineLayout that is compatible for set n, with the VkPipelineLayout used to create the current VkPipeline, as described in ???

For each push constant that is statically used by the VkPipeline bound to the pipeline bind point used by this command, a push constant value must have been set for the same pipeline bind point, with a VkPipelineLayout that is compatible for push constants, with the VkPipelineLayout used to create the current VkPipeline, as described in ???

Descriptors in each bound descriptor set, specified via vkCmdBindDescriptorSets, must be valid if they are statically used by the VkPipeline bound to the pipeline bind point used by this command

A valid pipeline must be bound to the pipeline bind point used by this command

If the VkPipeline object bound to the pipeline bind point used by this command requires any dynamic state, that state must have been set for commandBuffer

If the VkPipeline object bound to the pipeline bind point used by this command accesses a VkSampler object that uses unnormalized coordinates, that sampler must not be used to sample from any VkImage with a VkImageView of the type VK_IMAGE_VIEW_TYPE_3D, VK_IMAGE_VIEW_TYPE_CUBE, VK_IMAGE_VIEW_TYPE_1D_ARRAY, VK_IMAGE_VIEW_TYPE_2D_ARRAY or VK_IMAGE_VIEW_TYPE_CUBE_ARRAY, in any shader stage

If the VkPipeline object bound to the pipeline bind point used by this command accesses a VkSampler object that uses unnormalized coordinates, that sampler must not be used with any of the SPIR-V OpImageSample* or OpImageSparseSample* instructions with ImplicitLod, Dref or Proj in their name, in any shader stage

If the VkPipeline object bound to the pipeline bind point used by this command accesses a VkSampler object that uses unnormalized coordinates, that sampler must not be used with any of the SPIR-V OpImageSample* or OpImageSparseSample* instructions that includes a LOD bias or any offset values, in any shader stage

If the robust buffer access feature is not enabled, and if the VkPipeline object bound to the pipeline bind point used by this command accesses a uniform buffer, it must not access values outside of the range of the buffer as specified in the descriptor set bound to the same pipeline bind point

If the robust buffer access feature is not enabled, and if the VkPipeline object bound to the pipeline bind point used by this command accesses a storage buffer, it must not access values outside of the range of the buffer as specified in the descriptor set bound to the same pipeline bind point

baseGroupX must be less than VkPhysicalDeviceLimits::maxComputeWorkGroupCount[0]

baseGroupX must be less than VkPhysicalDeviceLimits::maxComputeWorkGroupCount[1]

baseGroupZ must be less than VkPhysicalDeviceLimits::maxComputeWorkGroupCount[2]

groupCountX must be less than or equal to VkPhysicalDeviceLimits::maxComputeWorkGroupCount[0] minus baseGroupX

groupCountY must be less than or equal to VkPhysicalDeviceLimits::maxComputeWorkGroupCount[1] minus baseGroupY

groupCountZ must be less than or equal to VkPhysicalDeviceLimits::maxComputeWorkGroupCount[2] minus baseGroupZ

If any of baseGroupX, baseGroupY, or baseGroupZ are not zero, then the bound compute pipeline must have been created with the VK_PIPELINE_CREATE_DISPATCH_BASE flag.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support compute operations

This command must only be called outside of a render pass instance

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Outside         | Compute         |                 |
| Secondary       |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer

cmdSetDeviceMask :: CommandBuffer -> Word32 -> IO ()

vkCmdSetDeviceMask - Modify device mask of a command buffer

Parameters

commandBuffer is command buffer whose current device mask is modified.

deviceMask is the new value of the current device mask.

Description

deviceMask is used to filter out subsequent commands from executing on all physical devices whose bit indices are not set in the mask, except commands beginning a render pass instance, commands transitioning to the next subpass in the render pass instance, and commands ending a render pass instance, which always execute on the set of physical devices whose bit indices are included in the deviceMask member of the instance of the VkDeviceGroupRenderPassBeginInfoKHR structure passed to the command beginning the corresponding render pass instance.

Valid Usage

deviceMask must be a valid device mask value

deviceMask must not be zero

deviceMask must not include any set bits that were not in the VkDeviceGroupCommandBufferBeginInfo::deviceMask value when the command buffer began recording.

If vkCmdSetDeviceMask is called inside a render pass instance, deviceMask must not include any set bits that were not in the VkDeviceGroupRenderPassBeginInfo::deviceMask value when the render pass instance began recording.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, compute, or transfer operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
|                 |                 | Transfer        |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer

getDeviceGroupPeerMemoryFeatures :: Device -> Word32 -> Word32 -> Word32 -> IO PeerMemoryFeatureFlags

vkGetDeviceGroupPeerMemoryFeatures - Query supported peer memory features of a device

Parameters

device is the logical device that owns the memory.

heapIndex is the index of the memory heap from which the memory is allocated.

localDeviceIndex is the device index of the physical device that performs the memory access.

remoteDeviceIndex is the device index of the physical device that the memory is allocated for.

pPeerMemoryFeatures is a pointer to a bitmask of VkPeerMemoryFeatureFlagBits indicating which types of memory accesses are supported for the combination of heap, local, and remote devices.

Valid Usage (Implicit)

See Also

VkDevice, VkPeerMemoryFeatureFlags

pattern STRUCTURE_TYPE_MEMORY_ALLOCATE_FLAGS_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEVICE_GROUP_RENDER_PASS_BEGIN_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEVICE_GROUP_COMMAND_BUFFER_BEGIN_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEVICE_GROUP_SUBMIT_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEVICE_GROUP_BIND_SPARSE_INFO :: a ~ StructureType => a

pattern PIPELINE_CREATE_VIEW_INDEX_FROM_DEVICE_INDEX_BIT :: a ~ PipelineCreateFlagBits => a

pattern PIPELINE_CREATE_DISPATCH_BASE :: a ~ PipelineCreateFlagBits => a

pattern DEPENDENCY_DEVICE_GROUP_BIT :: a ~ DependencyFlagBits => a