Graphics.Vulkan.Core11.Promoted_from_VK_KHR_external_fence

Documentation

withCStructExportFenceCreateInfo :: ExportFenceCreateInfo -> (VkExportFenceCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkExportFenceCreateInfo and marshal a ExportFenceCreateInfo into it. The VkExportFenceCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructExportFenceCreateInfo :: VkExportFenceCreateInfo -> IO ExportFenceCreateInfo

A function to read a VkExportFenceCreateInfo and all additional structures in the pointer chain into a ExportFenceCreateInfo.

data ExportFenceCreateInfo

VkExportFenceCreateInfo - Structure specifying handle types that can be exported from a fence

Valid Usage

The bits in handleTypes must be supported and compatible, as reported by VkExternalFenceProperties.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_EXPORT_FENCE_CREATE_INFO

handleTypes must be a valid combination of VkExternalFenceHandleTypeFlagBits values

See Also

VkExternalFenceHandleTypeFlags, VkStructureType

Constructors

ExportFenceCreateInfo

Fields

next :: Maybe SomeVkStruct

handleTypes :: ExternalFenceHandleTypeFlags

type FenceImportFlagBits = VkFenceImportFlagBits

VkFenceImportFlagBits - Bitmask specifying additional parameters of fence payload import

See Also

VkFenceImportFlags

pattern FENCE_IMPORT_TEMPORARY_BIT :: a ~ FenceImportFlagBits => a

VK_FENCE_IMPORT_TEMPORARY_BIT specifies that the fence payload will be imported only temporarily, as described in Importing Fence Payloads, regardless of the permanence of handleType.

type FenceImportFlagBitsKHR = FenceImportFlagBits

type FenceImportFlags = FenceImportFlagBits

VkFenceImportFlags - Bitmask of VkFenceImportFlagBits

Description

VkFenceImportFlags is a bitmask type for setting a mask of zero or more VkFenceImportFlagBits.

See Also

VkFenceImportFlagBits, VkImportFenceFdInfoKHR, VkImportFenceWin32HandleInfoKHR

type FenceImportFlagsKHR = FenceImportFlags

pattern STRUCTURE_TYPE_EXPORT_FENCE_CREATE_INFO :: a ~ StructureType => a