Graphics.Vulkan.Core11.Promoted_from_VK_KHR_external_fence_capabilities

Documentation

type ExternalFenceFeatureFlagBits = VkExternalFenceFeatureFlagBits

VkExternalFenceFeatureFlagBits - Bitfield describing features of an external fence handle type

See Also

VkExternalFenceFeatureFlags

pattern EXTERNAL_FENCE_FEATURE_EXPORTABLE_BIT :: a ~ ExternalFenceFeatureFlagBits => a

VK_EXTERNAL_FENCE_FEATURE_EXPORTABLE_BIT specifies handles of this type can be exported from Vulkan fence objects.

pattern EXTERNAL_FENCE_FEATURE_IMPORTABLE_BIT :: a ~ ExternalFenceFeatureFlagBits => a

VK_EXTERNAL_FENCE_FEATURE_IMPORTABLE_BIT specifies handles of this type can be imported to Vulkan fence objects.

type ExternalFenceFeatureFlagBitsKHR = ExternalFenceFeatureFlagBits

type ExternalFenceFeatureFlags = ExternalFenceFeatureFlagBits

VkExternalFenceFeatureFlags - Bitmask of VkExternalFenceFeatureFlagBits

Description

VkExternalFenceFeatureFlags is a bitmask type for setting a mask of zero or more VkExternalFenceFeatureFlagBits.

See Also

VkExternalFenceFeatureFlagBits, VkExternalFenceProperties

type ExternalFenceFeatureFlagsKHR = ExternalFenceFeatureFlags

type ExternalFenceHandleTypeFlagBits = VkExternalFenceHandleTypeFlagBits

VkExternalFenceHandleTypeFlagBits - Bitmask of valid external fence handle types

Description

Some external fence handle types can only be shared within the same underlying physical device and/or the same driver version, as defined in the following table:

+----------------------+----------------------+-----------------------+
| Handle type          | 'Graphics.Vulkan.C.C | 'Graphics.Vulkan.C.Co |
|                      | ore11.Promoted_from_ | re11.Promoted_from_VK |
|                      | VK_KHR_external_memo | _KHR_external_memory_ |
|                      | ry_capabilities.VkPh | capabilities.VkPhysic |
|                      | ysicalDeviceIDProper | alDeviceIDProperties' |
|                      | ties'::@driverUUID@  | ::@deviceUUID@        |
+----------------------+----------------------+-----------------------+
| 'Graphics.Vulkan.C.C | Must match           | Must match            |
| ore11.Promoted_from_ |                      |                       |
| VK_KHR_external_fenc |                      |                       |
| e_capabilities.VK_EX |                      |                       |
| TERNAL_FENCE_HANDLE_ |                      |                       |
| TYPE_OPAQUE_FD_BIT'  |                      |                       |
+----------------------+----------------------+-----------------------+
| 'Graphics.Vulkan.C.C | Must match           | Must match            |
| ore11.Promoted_from_ |                      |                       |
| VK_KHR_external_fenc |                      |                       |
| e_capabilities.VK_EX |                      |                       |
| TERNAL_FENCE_HANDLE_ |                      |                       |
| TYPE_OPAQUE_WIN32_BI |                      |                       |
| T'                   |                      |                       |
+----------------------+----------------------+-----------------------+
| 'Graphics.Vulkan.C.C | Must match           | Must match            |
| ore11.Promoted_from_ |                      |                       |
| VK_KHR_external_fenc |                      |                       |
| e_capabilities.VK_EX |                      |                       |
| TERNAL_FENCE_HANDLE_ |                      |                       |
| TYPE_OPAQUE_WIN32_KM |                      |                       |
| T_BIT'               |                      |                       |
+----------------------+----------------------+-----------------------+
| 'Graphics.Vulkan.C.C | No restriction       | No restriction        |
| ore11.Promoted_from_ |                      |                       |
| VK_KHR_external_fenc |                      |                       |
| e_capabilities.VK_EX |                      |                       |
| TERNAL_FENCE_HANDLE_ |                      |                       |
| TYPE_SYNC_FD_BIT'    |                      |                       |
+----------------------+----------------------+-----------------------+

External fence handle types compatibility

See Also

VkExternalFenceHandleTypeFlags, VkFenceGetFdInfoKHR, VkFenceGetWin32HandleInfoKHR, VkImportFenceFdInfoKHR, VkImportFenceWin32HandleInfoKHR, VkPhysicalDeviceExternalFenceInfo

pattern EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_FD_BIT :: a ~ ExternalFenceHandleTypeFlagBits => a

VK_EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_FD_BIT specifies a POSIX file descriptor handle that has only limited valid usage outside of Vulkan and other compatible APIs. It must be compatible with the POSIX system calls dup, dup2, close, and the non-standard system call dup3. Additionally, it must be transportable over a socket using an SCM_RIGHTS control message. It owns a reference to the underlying synchronization primitive represented by its Vulkan fence object.

pattern EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_BIT :: a ~ ExternalFenceHandleTypeFlagBits => a

VK_EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_BIT specifies an NT handle that has only limited valid usage outside of Vulkan and other compatible APIs. It must be compatible with the functions DuplicateHandle, CloseHandle, CompareObjectHandles, GetHandleInformation, and SetHandleInformation. It owns a reference to the underlying synchronization primitive represented by its Vulkan fence object.

pattern EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT :: a ~ ExternalFenceHandleTypeFlagBits => a

VK_EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT specifies a global share handle that has only limited valid usage outside of Vulkan and other compatible APIs. It is not compatible with any native APIs. It does not own a reference to the underlying synchronization primitive represented by its Vulkan fence object, and will therefore become invalid when all Vulkan fence objects associated with it are destroyed.

pattern EXTERNAL_FENCE_HANDLE_TYPE_SYNC_FD_BIT :: a ~ ExternalFenceHandleTypeFlagBits => a

VK_EXTERNAL_FENCE_HANDLE_TYPE_SYNC_FD_BIT specifies a POSIX file descriptor handle to a Linux Sync File or Android Fence. It can be used with any native API accepting a valid sync file or fence as input. It owns a reference to the underlying synchronization primitive associated with the file descriptor. Implementations which support importing this handle type must accept any type of sync or fence FD supported by the native system they are running on.

type ExternalFenceHandleTypeFlagBitsKHR = ExternalFenceHandleTypeFlagBits

type ExternalFenceHandleTypeFlags = ExternalFenceHandleTypeFlagBits

VkExternalFenceHandleTypeFlags - Bitmask of VkExternalFenceHandleTypeFlagBits

Description

VkExternalFenceHandleTypeFlags is a bitmask type for setting a mask of zero or more VkExternalFenceHandleTypeFlagBits.

See Also

VkExportFenceCreateInfo, VkExternalFenceHandleTypeFlagBits, VkExternalFenceProperties

type ExternalFenceHandleTypeFlagsKHR = ExternalFenceHandleTypeFlags

withCStructExternalFenceProperties :: ExternalFenceProperties -> (VkExternalFenceProperties -> IO a) -> IO a

A function to temporarily allocate memory for a VkExternalFenceProperties and marshal a ExternalFenceProperties into it. The VkExternalFenceProperties is only valid inside the provided computation and must not be returned out of it.

fromCStructExternalFenceProperties :: VkExternalFenceProperties -> IO ExternalFenceProperties

A function to read a VkExternalFenceProperties and all additional structures in the pointer chain into a ExternalFenceProperties.

data ExternalFenceProperties

VkExternalFenceProperties - Structure describing supported external fence handle features

Description

If handleType is not supported by the implementation, then VkExternalFenceProperties::externalFenceFeatures will be set to zero.

Valid Usage (Implicit)

See Also

VkExternalFenceFeatureFlags, VkExternalFenceHandleTypeFlags, VkStructureType, vkGetPhysicalDeviceExternalFenceProperties, vkGetPhysicalDeviceExternalFencePropertiesKHR

Constructors

ExternalFenceProperties

Fields

next :: Maybe SomeVkStruct

exportFromImportedHandleTypes :: ExternalFenceHandleTypeFlags

compatibleHandleTypes :: ExternalFenceHandleTypeFlags

externalFenceFeatures :: ExternalFenceFeatureFlags

withCStructPhysicalDeviceExternalFenceInfo :: PhysicalDeviceExternalFenceInfo -> (VkPhysicalDeviceExternalFenceInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceExternalFenceInfo and marshal a PhysicalDeviceExternalFenceInfo into it. The VkPhysicalDeviceExternalFenceInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceExternalFenceInfo :: VkPhysicalDeviceExternalFenceInfo -> IO PhysicalDeviceExternalFenceInfo

A function to read a VkPhysicalDeviceExternalFenceInfo and all additional structures in the pointer chain into a PhysicalDeviceExternalFenceInfo.

data PhysicalDeviceExternalFenceInfo

VkPhysicalDeviceExternalFenceInfo - Structure specifying fence creation parameters.

Description

Note

Handles of type VK_EXTERNAL_FENCE_HANDLE_TYPE_SYNC_FD_BIT generated by the implementation may represent either Linux Sync Files or Android Fences at the implementation’s discretion. Applications should only use operations defined for both types of file descriptors, unless they know via means external to Vulkan the type of the file descriptor, or are prepared to deal with the system-defined operation failures resulting from using the wrong type.

Valid Usage (Implicit)

See Also

VkExternalFenceHandleTypeFlagBits, VkStructureType, vkGetPhysicalDeviceExternalFenceProperties, vkGetPhysicalDeviceExternalFencePropertiesKHR

Constructors

PhysicalDeviceExternalFenceInfo

Fields

next :: Maybe SomeVkStruct

handleType :: ExternalFenceHandleTypeFlagBits

getPhysicalDeviceExternalFenceProperties :: PhysicalDevice -> PhysicalDeviceExternalFenceInfo -> IO ExternalFenceProperties

vkGetPhysicalDeviceExternalFenceProperties - Function for querying external fence handle capabilities.

Parameters

physicalDevice is the physical device from which to query the fence capabilities.

pExternalFenceInfo points to an instance of the VkPhysicalDeviceExternalFenceInfo structure, describing the parameters that would be consumed by vkCreateFence.

pExternalFenceProperties points to an instance of the VkExternalFenceProperties structure in which capabilities are returned.

Valid Usage (Implicit)

See Also

VkExternalFenceProperties, VkPhysicalDevice, VkPhysicalDeviceExternalFenceInfo

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_FENCE_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_EXTERNAL_FENCE_PROPERTIES :: a ~ StructureType => a