Graphics.Vulkan.Core11.Promoted_from_VK_KHR_external_memory

Documentation

withCStructExportMemoryAllocateInfo :: ExportMemoryAllocateInfo -> (VkExportMemoryAllocateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkExportMemoryAllocateInfo and marshal a ExportMemoryAllocateInfo into it. The VkExportMemoryAllocateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructExportMemoryAllocateInfo :: VkExportMemoryAllocateInfo -> IO ExportMemoryAllocateInfo

A function to read a VkExportMemoryAllocateInfo and all additional structures in the pointer chain into a ExportMemoryAllocateInfo.

data ExportMemoryAllocateInfo

VkExportMemoryAllocateInfo - Specify exportable handle types for a device memory object

Valid Usage

The bits in handleTypes must be supported and compatible, as reported by VkExternalImageFormatProperties or VkExternalBufferProperties.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_EXPORT_MEMORY_ALLOCATE_INFO

handleTypes must be a valid combination of VkExternalMemoryHandleTypeFlagBits values

See Also

VkExternalMemoryHandleTypeFlags, VkStructureType

Constructors

ExportMemoryAllocateInfo

Fields

next :: Maybe SomeVkStruct

handleTypes :: ExternalMemoryHandleTypeFlags

withCStructExternalMemoryBufferCreateInfo :: ExternalMemoryBufferCreateInfo -> (VkExternalMemoryBufferCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkExternalMemoryBufferCreateInfo and marshal a ExternalMemoryBufferCreateInfo into it. The VkExternalMemoryBufferCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructExternalMemoryBufferCreateInfo :: VkExternalMemoryBufferCreateInfo -> IO ExternalMemoryBufferCreateInfo

A function to read a VkExternalMemoryBufferCreateInfo and all additional structures in the pointer chain into a ExternalMemoryBufferCreateInfo.

data ExternalMemoryBufferCreateInfo

VkExternalMemoryBufferCreateInfo - Specify that a buffer may be backed by external memory

Valid Usage (Implicit)

See Also

VkExternalMemoryHandleTypeFlags, VkStructureType

Constructors

ExternalMemoryBufferCreateInfo

Fields

next :: Maybe SomeVkStruct

handleTypes :: ExternalMemoryHandleTypeFlags

withCStructExternalMemoryImageCreateInfo :: ExternalMemoryImageCreateInfo -> (VkExternalMemoryImageCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkExternalMemoryImageCreateInfo and marshal a ExternalMemoryImageCreateInfo into it. The VkExternalMemoryImageCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructExternalMemoryImageCreateInfo :: VkExternalMemoryImageCreateInfo -> IO ExternalMemoryImageCreateInfo

A function to read a VkExternalMemoryImageCreateInfo and all additional structures in the pointer chain into a ExternalMemoryImageCreateInfo.

data ExternalMemoryImageCreateInfo

VkExternalMemoryImageCreateInfo - Specify that an image may be backed by external memory

Valid Usage (Implicit)

See Also

VkExternalMemoryHandleTypeFlags, VkStructureType

Constructors

ExternalMemoryImageCreateInfo

Fields

next :: Maybe SomeVkStruct

handleTypes :: ExternalMemoryHandleTypeFlags

pattern STRUCTURE_TYPE_EXTERNAL_MEMORY_BUFFER_CREATE_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_EXTERNAL_MEMORY_IMAGE_CREATE_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_EXPORT_MEMORY_ALLOCATE_INFO :: a ~ StructureType => a

pattern ERROR_INVALID_EXTERNAL_HANDLE :: a ~ Result => a