Graphics.Vulkan.Core11.Promoted_from_VK_KHR_external_memory_capabilities

Documentation

withCStructExternalBufferProperties :: ExternalBufferProperties -> (VkExternalBufferProperties -> IO a) -> IO a

A function to temporarily allocate memory for a VkExternalBufferProperties and marshal a ExternalBufferProperties into it. The VkExternalBufferProperties is only valid inside the provided computation and must not be returned out of it.

fromCStructExternalBufferProperties :: VkExternalBufferProperties -> IO ExternalBufferProperties

A function to read a VkExternalBufferProperties and all additional structures in the pointer chain into a ExternalBufferProperties.

data ExternalBufferProperties

VkExternalBufferProperties - Structure specifying supported external handle capabilities

Valid Usage (Implicit)

See Also

VkExternalMemoryProperties, VkStructureType, vkGetPhysicalDeviceExternalBufferProperties, vkGetPhysicalDeviceExternalBufferPropertiesKHR

Constructors

ExternalBufferProperties

Fields

next :: Maybe SomeVkStruct

externalMemoryProperties :: ExternalMemoryProperties

withCStructExternalImageFormatProperties :: ExternalImageFormatProperties -> (VkExternalImageFormatProperties -> IO a) -> IO a

A function to temporarily allocate memory for a VkExternalImageFormatProperties and marshal a ExternalImageFormatProperties into it. The VkExternalImageFormatProperties is only valid inside the provided computation and must not be returned out of it.

fromCStructExternalImageFormatProperties :: VkExternalImageFormatProperties -> IO ExternalImageFormatProperties

A function to read a VkExternalImageFormatProperties and all additional structures in the pointer chain into a ExternalImageFormatProperties.

data ExternalImageFormatProperties

VkExternalImageFormatProperties - Structure specifying supported external handle properties

Valid Usage (Implicit)

See Also

VkExternalMemoryProperties, VkStructureType

Constructors

ExternalImageFormatProperties

Fields

next :: Maybe SomeVkStruct

externalMemoryProperties :: ExternalMemoryProperties

type ExternalMemoryFeatureFlagBits = VkExternalMemoryFeatureFlagBits

VkExternalMemoryFeatureFlagBits - Bitmask specifying features of an external memory handle type

Description

Because their semantics in external APIs roughly align with that of an image or buffer with a dedicated allocation in Vulkan, implementations are required to report VK_EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT for the following external handle types:

See Also

VkExternalMemoryFeatureFlags

pattern EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT :: a ~ ExternalMemoryFeatureFlagBits => a

VK_EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT specifies that images or buffers created with the specified parameters and handle type must use the mechanisms defined by VkMemoryDedicatedRequirements and VkMemoryDedicatedAllocateInfo to create (or import) a dedicated allocation for the image or buffer.

pattern EXTERNAL_MEMORY_FEATURE_EXPORTABLE_BIT :: a ~ ExternalMemoryFeatureFlagBits => a

VK_EXTERNAL_MEMORY_FEATURE_EXPORTABLE_BIT specifies that handles of this type can be exported from Vulkan memory objects.

pattern EXTERNAL_MEMORY_FEATURE_IMPORTABLE_BIT :: a ~ ExternalMemoryFeatureFlagBits => a

VK_EXTERNAL_MEMORY_FEATURE_IMPORTABLE_BIT specifies that handles of this type can be imported as Vulkan memory objects.

type ExternalMemoryFeatureFlagBitsKHR = ExternalMemoryFeatureFlagBits

type ExternalMemoryFeatureFlags = ExternalMemoryFeatureFlagBits

VkExternalMemoryFeatureFlags - Bitmask of VkExternalMemoryFeatureFlagBits

Description

VkExternalMemoryFeatureFlags is a bitmask type for setting a mask of zero or more VkExternalMemoryFeatureFlagBits.

See Also

VkExternalMemoryFeatureFlagBits, VkExternalMemoryProperties

type ExternalMemoryFeatureFlagsKHR = ExternalMemoryFeatureFlags

type ExternalMemoryHandleTypeFlagBits = VkExternalMemoryHandleTypeFlagBits

VkExternalMemoryHandleTypeFlagBits - Bit specifying external memory handle types

Description

Some external memory handle types can only be shared within the same underlying physical device and/or the same driver version, as defined in the following table:

+----------------------+----------------------+-----------------------+
| Handle type          | 'Graphics.Vulkan.C.C | 'Graphics.Vulkan.C.Co |
|                      | ore11.Promoted_from_ | re11.Promoted_from_VK |
|                      | VK_KHR_external_memo | _KHR_external_memory_ |
|                      | ry_capabilities.VkPh | capabilities.VkPhysic |
|                      | ysicalDeviceIDProper | alDeviceIDProperties' |
|                      | ties'::@driverUUID@  | ::@deviceUUID@        |
+----------------------+----------------------+-----------------------+
| 'Graphics.Vulkan.C.C | Must match           | Must match            |
| ore11.Promoted_from_ |                      |                       |
| VK_KHR_external_memo |                      |                       |
| ry_capabilities.VK_E |                      |                       |
| XTERNAL_MEMORY_HANDL |                      |                       |
| E_TYPE_OPAQUE_FD_BIT |                      |                       |
| '                    |                      |                       |
+----------------------+----------------------+-----------------------+
| 'Graphics.Vulkan.C.C | Must match           | Must match            |
| ore11.Promoted_from_ |                      |                       |
| VK_KHR_external_memo |                      |                       |
| ry_capabilities.VK_E |                      |                       |
| XTERNAL_MEMORY_HANDL |                      |                       |
| E_TYPE_OPAQUE_WIN32_ |                      |                       |
| BIT'                 |                      |                       |
+----------------------+----------------------+-----------------------+
| 'Graphics.Vulkan.C.C | Must match           | Must match            |
| ore11.Promoted_from_ |                      |                       |
| VK_KHR_external_memo |                      |                       |
| ry_capabilities.VK_E |                      |                       |
| XTERNAL_MEMORY_HANDL |                      |                       |
| E_TYPE_OPAQUE_WIN32_ |                      |                       |
| KMT_BIT'             |                      |                       |
+----------------------+----------------------+-----------------------+
| 'Graphics.Vulkan.C.C | Must match           | Must match            |
| ore11.Promoted_from_ |                      |                       |
| VK_KHR_external_memo |                      |                       |
| ry_capabilities.VK_E |                      |                       |
| XTERNAL_MEMORY_HANDL |                      |                       |
| E_TYPE_D3D11_TEXTURE |                      |                       |
| _BIT'                |                      |                       |
+----------------------+----------------------+-----------------------+
| 'Graphics.Vulkan.C.C | Must match           | Must match            |
| ore11.Promoted_from_ |                      |                       |
| VK_KHR_external_memo |                      |                       |
| ry_capabilities.VK_E |                      |                       |
| XTERNAL_MEMORY_HANDL |                      |                       |
| E_TYPE_D3D11_TEXTURE |                      |                       |
| _KMT_BIT'            |                      |                       |
+----------------------+----------------------+-----------------------+
| 'Graphics.Vulkan.C.C | Must match           | Must match            |
| ore11.Promoted_from_ |                      |                       |
| VK_KHR_external_memo |                      |                       |
| ry_capabilities.VK_E |                      |                       |
| XTERNAL_MEMORY_HANDL |                      |                       |
| E_TYPE_D3D12_HEAP_BI |                      |                       |
| T'                   |                      |                       |
+----------------------+----------------------+-----------------------+
| 'Graphics.Vulkan.C.C | Must match           | Must match            |
| ore11.Promoted_from_ |                      |                       |
| VK_KHR_external_memo |                      |                       |
| ry_capabilities.VK_E |                      |                       |
| XTERNAL_MEMORY_HANDL |                      |                       |
| E_TYPE_D3D12_RESOURC |                      |                       |
| E_BIT'               |                      |                       |
+----------------------+----------------------+-----------------------+

External memory handle types compatibility

See Also

VkExternalMemoryHandleTypeFlags, VkImportMemoryFdInfoKHR, VkImportMemoryHostPointerInfoEXT, VkImportMemoryWin32HandleInfoKHR, VkMemoryGetFdInfoKHR, VkMemoryGetWin32HandleInfoKHR, VkPhysicalDeviceExternalBufferInfo, VkPhysicalDeviceExternalImageFormatInfo, vkGetMemoryFdPropertiesKHR, vkGetMemoryHostPointerPropertiesEXT, vkGetMemoryWin32HandlePropertiesKHR

pattern EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_FD_BIT :: a ~ ExternalMemoryHandleTypeFlagBits => a

VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_FD_BIT specifies a POSIX file descriptor handle that has only limited valid usage outside of Vulkan and other compatible APIs. It must be compatible with the POSIX system calls dup, dup2, close, and the non-standard system call dup3. Additionally, it must be transportable over a socket using an SCM_RIGHTS control message. It owns a reference to the underlying memory resource represented by its Vulkan memory object.

pattern EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT :: a ~ ExternalMemoryHandleTypeFlagBits => a

VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT specifies an NT handle that has only limited valid usage outside of Vulkan and other compatible APIs. It must be compatible with the functions DuplicateHandle, CloseHandle, CompareObjectHandles, GetHandleInformation, and SetHandleInformation. It owns a reference to the underlying memory resource represented by its Vulkan memory object.

pattern EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT :: a ~ ExternalMemoryHandleTypeFlagBits => a

VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT specifies a global share handle that has only limited valid usage outside of Vulkan and other compatible APIs. It is not compatible with any native APIs. It does not own a reference to the underlying memory resource represented its Vulkan memory object, and will therefore become invalid when all Vulkan memory objects associated with it are destroyed.

pattern EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_BIT :: a ~ ExternalMemoryHandleTypeFlagBits => a

VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_BIT specifies an NT handle returned by IDXGIResource1::CreateSharedHandle referring to a Direct3D 10 or 11 texture resource. It owns a reference to the memory used by the Direct3D resource.

pattern EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_KMT_BIT :: a ~ ExternalMemoryHandleTypeFlagBits => a

VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_KMT_BIT specifies a global share handle returned by IDXGIResource::GetSharedHandle referring to a Direct3D 10 or 11 texture resource. It does not own a reference to the underlying Direct3D resource, and will therefore become invalid when all Vulkan memory objects and Direct3D resources associated with it are destroyed.

pattern EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_HEAP_BIT :: a ~ ExternalMemoryHandleTypeFlagBits => a

VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_HEAP_BIT specifies an NT handle returned by ID3D12Device::CreateSharedHandle referring to a Direct3D 12 heap resource. It owns a reference to the resources used by the Direct3D heap.

pattern EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_RESOURCE_BIT :: a ~ ExternalMemoryHandleTypeFlagBits => a

VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_RESOURCE_BIT specifies an NT handle returned by ID3D12Device::CreateSharedHandle referring to a Direct3D 12 committed resource. It owns a reference to the memory used by the Direct3D resource.

pattern EXTERNAL_MEMORY_HANDLE_TYPE_DMA_BUF_BIT_EXT :: a ~ ExternalMemoryHandleTypeFlagBits => a

pattern EXTERNAL_MEMORY_HANDLE_TYPE_ANDROID_HARDWARE_BUFFER_BIT_ANDROID :: a ~ ExternalMemoryHandleTypeFlagBits => a

pattern EXTERNAL_MEMORY_HANDLE_TYPE_HOST_ALLOCATION_BIT_EXT :: a ~ ExternalMemoryHandleTypeFlagBits => a

pattern EXTERNAL_MEMORY_HANDLE_TYPE_HOST_MAPPED_FOREIGN_MEMORY_BIT_EXT :: a ~ ExternalMemoryHandleTypeFlagBits => a

type ExternalMemoryHandleTypeFlagBitsKHR = ExternalMemoryHandleTypeFlagBits

type ExternalMemoryHandleTypeFlags = ExternalMemoryHandleTypeFlagBits

VkExternalMemoryHandleTypeFlags - Bitmask of VkExternalMemoryHandleTypeFlagBits

Description

VkExternalMemoryHandleTypeFlags is a bitmask type for setting a mask of zero or more VkExternalMemoryHandleTypeFlagBits.

See Also

VkExportMemoryAllocateInfo, VkExternalMemoryBufferCreateInfo, VkExternalMemoryHandleTypeFlagBits, VkExternalMemoryImageCreateInfo, VkExternalMemoryProperties

type ExternalMemoryHandleTypeFlagsKHR = ExternalMemoryHandleTypeFlags

withCStructExternalMemoryProperties :: ExternalMemoryProperties -> (VkExternalMemoryProperties -> IO a) -> IO a

A function to temporarily allocate memory for a VkExternalMemoryProperties and marshal a ExternalMemoryProperties into it. The VkExternalMemoryProperties is only valid inside the provided computation and must not be returned out of it.

fromCStructExternalMemoryProperties :: VkExternalMemoryProperties -> IO ExternalMemoryProperties

A function to read a VkExternalMemoryProperties and all additional structures in the pointer chain into a ExternalMemoryProperties.

data ExternalMemoryProperties

VkExternalMemoryProperties - Structure specifying external memory handle type capabilities

Description

compatibleHandleTypes must include at least handleType. Inclusion of a handle type in compatibleHandleTypes does not imply the values returned in VkImageFormatProperties2 will be the same when VkPhysicalDeviceExternalImageFormatInfo::handleType is set to that type. The application is responsible for querying the capabilities of all handle types intended for concurrent use in a single image and intersecting them to obtain the compatible set of capabilities.

See Also

VkExternalBufferProperties, VkExternalImageFormatProperties, VkExternalMemoryFeatureFlags, VkExternalMemoryHandleTypeFlags

Constructors

ExternalMemoryProperties

Fields

externalMemoryFeatures :: ExternalMemoryFeatureFlags

exportFromImportedHandleTypes :: ExternalMemoryHandleTypeFlags

compatibleHandleTypes :: ExternalMemoryHandleTypeFlags

withCStructPhysicalDeviceExternalBufferInfo :: PhysicalDeviceExternalBufferInfo -> (VkPhysicalDeviceExternalBufferInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceExternalBufferInfo and marshal a PhysicalDeviceExternalBufferInfo into it. The VkPhysicalDeviceExternalBufferInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceExternalBufferInfo :: VkPhysicalDeviceExternalBufferInfo -> IO PhysicalDeviceExternalBufferInfo

A function to read a VkPhysicalDeviceExternalBufferInfo and all additional structures in the pointer chain into a PhysicalDeviceExternalBufferInfo.

data PhysicalDeviceExternalBufferInfo

VkPhysicalDeviceExternalBufferInfo - Structure specifying buffer creation parameters

Valid Usage (Implicit)

See Also

VkBufferCreateFlags, VkBufferUsageFlags, VkExternalMemoryHandleTypeFlagBits, VkStructureType, vkGetPhysicalDeviceExternalBufferProperties, vkGetPhysicalDeviceExternalBufferPropertiesKHR

Constructors

PhysicalDeviceExternalBufferInfo

Fields

next :: Maybe SomeVkStruct

flags :: BufferCreateFlags

usage :: BufferUsageFlags

handleType :: ExternalMemoryHandleTypeFlagBits

withCStructPhysicalDeviceExternalImageFormatInfo :: PhysicalDeviceExternalImageFormatInfo -> (VkPhysicalDeviceExternalImageFormatInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceExternalImageFormatInfo and marshal a PhysicalDeviceExternalImageFormatInfo into it. The VkPhysicalDeviceExternalImageFormatInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceExternalImageFormatInfo :: VkPhysicalDeviceExternalImageFormatInfo -> IO PhysicalDeviceExternalImageFormatInfo

A function to read a VkPhysicalDeviceExternalImageFormatInfo and all additional structures in the pointer chain into a PhysicalDeviceExternalImageFormatInfo.

data PhysicalDeviceExternalImageFormatInfo

VkPhysicalDeviceExternalImageFormatInfo - Structure specifying external image creation parameters

Description

If handleType is 0, vkGetPhysicalDeviceImageFormatProperties2 will behave as if VkPhysicalDeviceExternalImageFormatInfo was not present, and VkExternalImageFormatProperties will be ignored.

If handleType is not compatible with the format, type, tiling, usage, and flags specified in VkPhysicalDeviceImageFormatInfo2, then vkGetPhysicalDeviceImageFormatProperties2 returns VK_ERROR_FORMAT_NOT_SUPPORTED.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_IMAGE_FORMAT_INFO

If handleType is not 0, handleType must be a valid VkExternalMemoryHandleTypeFlagBits value

See Also

VkExternalMemoryHandleTypeFlagBits, VkStructureType

Constructors

PhysicalDeviceExternalImageFormatInfo

Fields

next :: Maybe SomeVkStruct

handleType :: ExternalMemoryHandleTypeFlagBits

withCStructPhysicalDeviceIDProperties :: PhysicalDeviceIDProperties -> (VkPhysicalDeviceIDProperties -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceIDProperties and marshal a PhysicalDeviceIDProperties into it. The VkPhysicalDeviceIDProperties is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceIDProperties :: VkPhysicalDeviceIDProperties -> IO PhysicalDeviceIDProperties

A function to read a VkPhysicalDeviceIDProperties and all additional structures in the pointer chain into a PhysicalDeviceIDProperties.

data PhysicalDeviceIDProperties

VkPhysicalDeviceIDProperties - Structure specifying IDs related to the physical device

Description

deviceUUID must be immutable for a given device across instances, processes, driver APIs, driver versions, and system reboots.

Applications can compare the driverUUID value across instance and process boundaries, and can make similar queries in external APIs to determine whether they are capable of sharing memory objects and resources using them with the device.

deviceUUID and/or driverUUID must be used to determine whether a particular external object can be shared between driver components, where such a restriction exists as defined in the compatibility table for the particular object type:

If deviceLUIDValid is VK_FALSE, the values of deviceLUID and deviceNodeMask are undefined. If deviceLUIDValid is VK_TRUE and Vulkan is running on the Windows operating system, the contents of deviceLUID can be cast to an LUID object and must be equal to the locally unique identifier of a IDXGIAdapter1 object that corresponds to physicalDevice. If deviceLUIDValid is VK_TRUE, deviceNodeMask must contain exactly one bit. If Vulkan is running on an operating system that supports the Direct3D 12 API and physicalDevice corresponds to an individual device in a linked device adapter, deviceNodeMask identifies the Direct3D 12 node corresponding to physicalDevice. Otherwise, deviceNodeMask must be 1.

Note

Although they have identical descriptions, VkPhysicalDeviceIDProperties::deviceUUID may differ from VkPhysicalDeviceProperties2::pipelineCacheUUID. The former is intended to identify and correlate devices across API and driver boundaries, while the latter is used to identify a compatible device and driver combination to use when serializing and de-serializing pipeline state.

Note

While VkPhysicalDeviceIDProperties::deviceUUID is specified to remain consistent across driver versions and system reboots, it is not intended to be usable as a serializable persistent identifier for a device. It may change when a device is physically added to, removed from, or moved to a different connector in a system while that system is powered down. Further, there is no reasonable way to verify with conformance testing that a given device retains the same UUID in a given system across all driver versions supported in that system. While implementations should make every effort to report consistent device UUIDs across driver versions, applications should avoid relying on the persistence of this value for uses other than identifying compatible devices for external object sharing purposes.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceIDProperties

Fields

next :: Maybe SomeVkStruct

deviceUUID :: ByteString

driverUUID :: ByteString

deviceLUID :: ByteString

deviceNodeMask :: Word32

deviceLUIDValid :: Bool

getPhysicalDeviceExternalBufferProperties :: PhysicalDevice -> PhysicalDeviceExternalBufferInfo -> IO ExternalBufferProperties

vkGetPhysicalDeviceExternalBufferProperties - Query external handle types supported by buffers

Parameters

physicalDevice is the physical device from which to query the buffer capabilities.

pExternalBufferInfo points to an instance of the VkPhysicalDeviceExternalBufferInfo structure, describing the parameters that would be consumed by vkCreateBuffer.

pExternalBufferProperties points to an instance of the VkExternalBufferProperties structure in which capabilities are returned.

Valid Usage (Implicit)

See Also

VkExternalBufferProperties, VkPhysicalDevice, VkPhysicalDeviceExternalBufferInfo

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_IMAGE_FORMAT_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_EXTERNAL_IMAGE_FORMAT_PROPERTIES :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_BUFFER_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_EXTERNAL_BUFFER_PROPERTIES :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_ID_PROPERTIES :: a ~ StructureType => a