Graphics.Vulkan.Core11.Promoted_from_VK_KHR_external_semaphore

Documentation

withCStructExportSemaphoreCreateInfo :: ExportSemaphoreCreateInfo -> (VkExportSemaphoreCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkExportSemaphoreCreateInfo and marshal a ExportSemaphoreCreateInfo into it. The VkExportSemaphoreCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructExportSemaphoreCreateInfo :: VkExportSemaphoreCreateInfo -> IO ExportSemaphoreCreateInfo

A function to read a VkExportSemaphoreCreateInfo and all additional structures in the pointer chain into a ExportSemaphoreCreateInfo.

data ExportSemaphoreCreateInfo

VkExportSemaphoreCreateInfo - Structure specifying handle types that can be exported from a semaphore

Valid Usage

The bits in handleTypes must be supported and compatible, as reported by VkExternalSemaphoreProperties.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_EXPORT_SEMAPHORE_CREATE_INFO

handleTypes must be a valid combination of VkExternalSemaphoreHandleTypeFlagBits values

See Also

VkExternalSemaphoreHandleTypeFlags, VkStructureType

Constructors

ExportSemaphoreCreateInfo

Fields

next :: Maybe SomeVkStruct

handleTypes :: ExternalSemaphoreHandleTypeFlags

type SemaphoreImportFlagBits = VkSemaphoreImportFlagBits

VkSemaphoreImportFlagBits - Bitmask specifying additional parameters of semaphore payload import

Description

These bits have the following meanings:

See Also

VkSemaphoreImportFlags

pattern SEMAPHORE_IMPORT_TEMPORARY_BIT :: a ~ SemaphoreImportFlagBits => a

VK_SEMAPHORE_IMPORT_TEMPORARY_BIT specifies that the semaphore payload will be imported only temporarily, as described in Importing Semaphore Payloads, regardless of the permanence of handleType.

type SemaphoreImportFlagBitsKHR = SemaphoreImportFlagBits

type SemaphoreImportFlags = SemaphoreImportFlagBits

VkSemaphoreImportFlags - Bitmask of VkSemaphoreImportFlagBits

Description

VkSemaphoreImportFlags is a bitmask type for setting a mask of zero or more VkSemaphoreImportFlagBits.

See Also

VkImportSemaphoreFdInfoKHR, VkImportSemaphoreWin32HandleInfoKHR, VkSemaphoreImportFlagBits

type SemaphoreImportFlagsKHR = SemaphoreImportFlags

pattern STRUCTURE_TYPE_EXPORT_SEMAPHORE_CREATE_INFO :: a ~ StructureType => a