Graphics.Vulkan.Core11.Promoted_from_VK_KHR_get_memory_requirements2

Documentation

withCStructBufferMemoryRequirementsInfo2 :: BufferMemoryRequirementsInfo2 -> (VkBufferMemoryRequirementsInfo2 -> IO a) -> IO a

A function to temporarily allocate memory for a VkBufferMemoryRequirementsInfo2 and marshal a BufferMemoryRequirementsInfo2 into it. The VkBufferMemoryRequirementsInfo2 is only valid inside the provided computation and must not be returned out of it.

fromCStructBufferMemoryRequirementsInfo2 :: VkBufferMemoryRequirementsInfo2 -> IO BufferMemoryRequirementsInfo2

A function to read a VkBufferMemoryRequirementsInfo2 and all additional structures in the pointer chain into a BufferMemoryRequirementsInfo2.

data BufferMemoryRequirementsInfo2

VkBufferMemoryRequirementsInfo2 - (None)

Valid Usage (Implicit)

See Also

VkBuffer, VkStructureType, vkGetBufferMemoryRequirements2, vkGetBufferMemoryRequirements2KHR

Constructors

BufferMemoryRequirementsInfo2

Fields

next :: Maybe SomeVkStruct

buffer :: Buffer

withCStructImageMemoryRequirementsInfo2 :: ImageMemoryRequirementsInfo2 -> (VkImageMemoryRequirementsInfo2 -> IO a) -> IO a

A function to temporarily allocate memory for a VkImageMemoryRequirementsInfo2 and marshal a ImageMemoryRequirementsInfo2 into it. The VkImageMemoryRequirementsInfo2 is only valid inside the provided computation and must not be returned out of it.

fromCStructImageMemoryRequirementsInfo2 :: VkImageMemoryRequirementsInfo2 -> IO ImageMemoryRequirementsInfo2

A function to read a VkImageMemoryRequirementsInfo2 and all additional structures in the pointer chain into a ImageMemoryRequirementsInfo2.

data ImageMemoryRequirementsInfo2

VkImageMemoryRequirementsInfo2 - (None)

Valid Usage (Implicit)

See Also

VkImage, VkStructureType, vkGetImageMemoryRequirements2, vkGetImageMemoryRequirements2KHR

Constructors

ImageMemoryRequirementsInfo2

Fields

next :: Maybe SomeVkStruct

image :: Image

withCStructImageSparseMemoryRequirementsInfo2 :: ImageSparseMemoryRequirementsInfo2 -> (VkImageSparseMemoryRequirementsInfo2 -> IO a) -> IO a

A function to temporarily allocate memory for a VkImageSparseMemoryRequirementsInfo2 and marshal a ImageSparseMemoryRequirementsInfo2 into it. The VkImageSparseMemoryRequirementsInfo2 is only valid inside the provided computation and must not be returned out of it.

fromCStructImageSparseMemoryRequirementsInfo2 :: VkImageSparseMemoryRequirementsInfo2 -> IO ImageSparseMemoryRequirementsInfo2

A function to read a VkImageSparseMemoryRequirementsInfo2 and all additional structures in the pointer chain into a ImageSparseMemoryRequirementsInfo2.

data ImageSparseMemoryRequirementsInfo2

VkImageSparseMemoryRequirementsInfo2 - (None)

Valid Usage (Implicit)

See Also

VkImage, VkStructureType, vkGetImageSparseMemoryRequirements2, vkGetImageSparseMemoryRequirements2KHR

Constructors

ImageSparseMemoryRequirementsInfo2

Fields

next :: Maybe SomeVkStruct

image :: Image

withCStructMemoryRequirements2 :: MemoryRequirements2 -> (VkMemoryRequirements2 -> IO a) -> IO a

A function to temporarily allocate memory for a VkMemoryRequirements2 and marshal a MemoryRequirements2 into it. The VkMemoryRequirements2 is only valid inside the provided computation and must not be returned out of it.

fromCStructMemoryRequirements2 :: VkMemoryRequirements2 -> IO MemoryRequirements2

A function to read a VkMemoryRequirements2 and all additional structures in the pointer chain into a MemoryRequirements2.

data MemoryRequirements2

VkMemoryRequirements2 - Structure specifying memory requirements

Valid Usage (Implicit)

See Also

VkMemoryRequirements, VkStructureType, vkGetBufferMemoryRequirements2, vkGetBufferMemoryRequirements2KHR, vkGetImageMemoryRequirements2, vkGetImageMemoryRequirements2KHR

Constructors

MemoryRequirements2

Fields

next :: Maybe SomeVkStruct

memoryRequirements :: MemoryRequirements

type MemoryRequirements2KHR = MemoryRequirements2

withCStructSparseImageMemoryRequirements2 :: SparseImageMemoryRequirements2 -> (VkSparseImageMemoryRequirements2 -> IO a) -> IO a

A function to temporarily allocate memory for a VkSparseImageMemoryRequirements2 and marshal a SparseImageMemoryRequirements2 into it. The VkSparseImageMemoryRequirements2 is only valid inside the provided computation and must not be returned out of it.

fromCStructSparseImageMemoryRequirements2 :: VkSparseImageMemoryRequirements2 -> IO SparseImageMemoryRequirements2

A function to read a VkSparseImageMemoryRequirements2 and all additional structures in the pointer chain into a SparseImageMemoryRequirements2.

data SparseImageMemoryRequirements2

VkSparseImageMemoryRequirements2 - (None)

Valid Usage (Implicit)

See Also

VkSparseImageMemoryRequirements, VkStructureType, vkGetImageSparseMemoryRequirements2, vkGetImageSparseMemoryRequirements2KHR

Constructors

SparseImageMemoryRequirements2

Fields

next :: Maybe SomeVkStruct

memoryRequirements :: SparseImageMemoryRequirements

getBufferMemoryRequirements2 :: Device -> BufferMemoryRequirementsInfo2 -> IO MemoryRequirements2

vkGetBufferMemoryRequirements2 - Returns the memory requirements for specified Vulkan object

Parameters

device is the logical device that owns the buffer.

pInfo is a pointer to an instance of the VkBufferMemoryRequirementsInfo2 structure containing parameters required for the memory requirements query.

pMemoryRequirements points to an instance of the VkMemoryRequirements2 structure in which the memory requirements of the buffer object are returned.

Valid Usage (Implicit)

See Also

VkBufferMemoryRequirementsInfo2, VkDevice, VkMemoryRequirements2

getImageMemoryRequirements2 :: Device -> ImageMemoryRequirementsInfo2 -> IO MemoryRequirements2

vkGetImageMemoryRequirements2 - Returns the memory requirements for specified Vulkan object

Parameters

device is the logical device that owns the image.

pInfo is a pointer to an instance of the VkImageMemoryRequirementsInfo2 structure containing parameters required for the memory requirements query.

pMemoryRequirements points to an instance of the VkMemoryRequirements2 structure in which the memory requirements of the image object are returned.

Valid Usage (Implicit)

See Also

VkDevice, VkImageMemoryRequirementsInfo2, VkMemoryRequirements2

getNumImageSparseMemoryRequirements2 :: Device -> ImageSparseMemoryRequirementsInfo2 -> IO Word32

vkGetImageSparseMemoryRequirements2 - Query the memory requirements for a sparse image

Parameters

device is the logical device that owns the image.

pInfo is a pointer to an instance of the VkImageSparseMemoryRequirementsInfo2 structure containing parameters required for the memory requirements query.

pSparseMemoryRequirementCount is a pointer to an integer related to the number of sparse memory requirements available or queried, as described below.

pSparseMemoryRequirements is either NULL or a pointer to an array of VkSparseImageMemoryRequirements2 structures.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pInfo must be a valid pointer to a valid VkImageSparseMemoryRequirementsInfo2 structure

pSparseMemoryRequirementCount must be a valid pointer to a uint32_t value

If the value referenced by pSparseMemoryRequirementCount is not 0, and pSparseMemoryRequirements is not NULL, pSparseMemoryRequirements must be a valid pointer to an array of pSparseMemoryRequirementCount VkSparseImageMemoryRequirements2 structures

See Also

VkDevice, VkImageSparseMemoryRequirementsInfo2, VkSparseImageMemoryRequirements2

getImageSparseMemoryRequirements2 :: Device -> ImageSparseMemoryRequirementsInfo2 -> Word32 -> IO (Vector SparseImageMemoryRequirements2)

vkGetImageSparseMemoryRequirements2 - Query the memory requirements for a sparse image

Parameters

device is the logical device that owns the image.

pInfo is a pointer to an instance of the VkImageSparseMemoryRequirementsInfo2 structure containing parameters required for the memory requirements query.

pSparseMemoryRequirementCount is a pointer to an integer related to the number of sparse memory requirements available or queried, as described below.

pSparseMemoryRequirements is either NULL or a pointer to an array of VkSparseImageMemoryRequirements2 structures.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pInfo must be a valid pointer to a valid VkImageSparseMemoryRequirementsInfo2 structure

pSparseMemoryRequirementCount must be a valid pointer to a uint32_t value

If the value referenced by pSparseMemoryRequirementCount is not 0, and pSparseMemoryRequirements is not NULL, pSparseMemoryRequirements must be a valid pointer to an array of pSparseMemoryRequirementCount VkSparseImageMemoryRequirements2 structures

See Also

VkDevice, VkImageSparseMemoryRequirementsInfo2, VkSparseImageMemoryRequirements2

getAllImageSparseMemoryRequirements2 :: Device -> ImageSparseMemoryRequirementsInfo2 -> IO (Vector SparseImageMemoryRequirements2)

Returns all the values available from getImageSparseMemoryRequirements2.

pattern STRUCTURE_TYPE_BUFFER_MEMORY_REQUIREMENTS_INFO_2 :: a ~ StructureType => a

pattern STRUCTURE_TYPE_IMAGE_MEMORY_REQUIREMENTS_INFO_2 :: a ~ StructureType => a

pattern STRUCTURE_TYPE_IMAGE_SPARSE_MEMORY_REQUIREMENTS_INFO_2 :: a ~ StructureType => a

pattern STRUCTURE_TYPE_MEMORY_REQUIREMENTS_2 :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SPARSE_IMAGE_MEMORY_REQUIREMENTS_2 :: a ~ StructureType => a