Graphics.Vulkan.Core11.Promoted_from_VK_KHR_get_physical_device_properties2

Documentation

withCStructFormatProperties2 :: FormatProperties2 -> (VkFormatProperties2 -> IO a) -> IO a

A function to temporarily allocate memory for a VkFormatProperties2 and marshal a FormatProperties2 into it. The VkFormatProperties2 is only valid inside the provided computation and must not be returned out of it.

fromCStructFormatProperties2 :: VkFormatProperties2 -> IO FormatProperties2

A function to read a VkFormatProperties2 and all additional structures in the pointer chain into a FormatProperties2.

data FormatProperties2

VkFormatProperties2 - Structure specifying image format properties

Valid Usage (Implicit)

See Also

VkFormatProperties, VkStructureType, vkGetPhysicalDeviceFormatProperties2, vkGetPhysicalDeviceFormatProperties2KHR

Constructors

FormatProperties2

Fields

next :: Maybe SomeVkStruct

formatProperties :: FormatProperties

withCStructImageFormatProperties2 :: ImageFormatProperties2 -> (VkImageFormatProperties2 -> IO a) -> IO a

A function to temporarily allocate memory for a VkImageFormatProperties2 and marshal a ImageFormatProperties2 into it. The VkImageFormatProperties2 is only valid inside the provided computation and must not be returned out of it.

fromCStructImageFormatProperties2 :: VkImageFormatProperties2 -> IO ImageFormatProperties2

A function to read a VkImageFormatProperties2 and all additional structures in the pointer chain into a ImageFormatProperties2.

data ImageFormatProperties2

VkImageFormatProperties2 - Structure specifying an image format properties

Description

If the combination of parameters to vkGetPhysicalDeviceImageFormatProperties2 is not supported by the implementation for use in vkCreateImage, then all members of imageFormatProperties will be filled with zero.

Note

Filling imageFormatProperties with zero for unsupported formats is an exception to the usual rule that output structures have undefined contents on error. This exception was unintentional, but is preserved for backwards compatibility. This exeption only applies to imageFormatProperties, not sType, pNext, or any structures chained from pNext.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMAGE_FORMAT_PROPERTIES_2

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkAndroidHardwareBufferUsageANDROID, VkExternalImageFormatProperties, VkFilterCubicImageViewImageFormatPropertiesEXT, VkSamplerYcbcrConversionImageFormatProperties, or VkTextureLODGatherFormatPropertiesAMD

Each sType member in the pNext chain must be unique

See Also

VkImageFormatProperties, VkStructureType, vkGetPhysicalDeviceImageFormatProperties2, vkGetPhysicalDeviceImageFormatProperties2KHR

Constructors

ImageFormatProperties2

Fields

next :: Maybe SomeVkStruct

imageFormatProperties :: ImageFormatProperties

withCStructPhysicalDeviceFeatures2 :: PhysicalDeviceFeatures2 -> (VkPhysicalDeviceFeatures2 -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceFeatures2 and marshal a PhysicalDeviceFeatures2 into it. The VkPhysicalDeviceFeatures2 is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceFeatures2 :: VkPhysicalDeviceFeatures2 -> IO PhysicalDeviceFeatures2

A function to read a VkPhysicalDeviceFeatures2 and all additional structures in the pointer chain into a PhysicalDeviceFeatures2.

data PhysicalDeviceFeatures2

VkPhysicalDeviceFeatures2 - Structure describing the fine-grained features that can be supported by an implementation

Members

The VkPhysicalDeviceFeatures2 structure is defined as:

Description

The pNext chain of this structure is used to extend the structure with features defined by extensions. This structure can be used in vkGetPhysicalDeviceFeatures2 or can be in the pNext chain of a VkDeviceCreateInfo structure, in which case it controls which features are enabled in the device in lieu of pEnabledFeatures.

Valid Usage (Implicit)

See Also

VkPhysicalDeviceFeatures, VkStructureType, vkGetPhysicalDeviceFeatures2, vkGetPhysicalDeviceFeatures2KHR

Constructors

PhysicalDeviceFeatures2

Fields

next :: Maybe SomeVkStruct

features :: PhysicalDeviceFeatures

withCStructPhysicalDeviceImageFormatInfo2 :: PhysicalDeviceImageFormatInfo2 -> (VkPhysicalDeviceImageFormatInfo2 -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceImageFormatInfo2 and marshal a PhysicalDeviceImageFormatInfo2 into it. The VkPhysicalDeviceImageFormatInfo2 is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceImageFormatInfo2 :: VkPhysicalDeviceImageFormatInfo2 -> IO PhysicalDeviceImageFormatInfo2

A function to read a VkPhysicalDeviceImageFormatInfo2 and all additional structures in the pointer chain into a PhysicalDeviceImageFormatInfo2.

data PhysicalDeviceImageFormatInfo2

VkPhysicalDeviceImageFormatInfo2 - Structure specifying image creation parameters

Description

The members of VkPhysicalDeviceImageFormatInfo2 correspond to the arguments to vkGetPhysicalDeviceImageFormatProperties, with sType and pNext added for extensibility.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGE_FORMAT_INFO_2

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkImageFormatListCreateInfoKHR, VkImageStencilUsageCreateInfoEXT, VkPhysicalDeviceExternalImageFormatInfo, VkPhysicalDeviceImageDrmFormatModifierInfoEXT, or VkPhysicalDeviceImageViewImageFormatInfoEXT

Each sType member in the pNext chain must be unique

format must be a valid VkFormat value

type must be a valid VkImageType value

tiling must be a valid VkImageTiling value

usage must be a valid combination of VkImageUsageFlagBits values

usage must not be 0

flags must be a valid combination of VkImageCreateFlagBits values

See Also

VkFormat, VkImageCreateFlags, VkImageTiling, VkImageType, VkImageUsageFlags, VkStructureType, vkGetPhysicalDeviceImageFormatProperties2, vkGetPhysicalDeviceImageFormatProperties2KHR

Constructors

PhysicalDeviceImageFormatInfo2

Fields

next :: Maybe SomeVkStruct

format :: Format

type' :: ImageType

tiling :: ImageTiling

usage :: ImageUsageFlags

flags :: ImageCreateFlags

withCStructPhysicalDeviceMemoryProperties2 :: PhysicalDeviceMemoryProperties2 -> (VkPhysicalDeviceMemoryProperties2 -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceMemoryProperties2 and marshal a PhysicalDeviceMemoryProperties2 into it. The VkPhysicalDeviceMemoryProperties2 is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceMemoryProperties2 :: VkPhysicalDeviceMemoryProperties2 -> IO PhysicalDeviceMemoryProperties2

A function to read a VkPhysicalDeviceMemoryProperties2 and all additional structures in the pointer chain into a PhysicalDeviceMemoryProperties2.

data PhysicalDeviceMemoryProperties2

VkPhysicalDeviceMemoryProperties2 - Structure specifying physical device memory properties

Valid Usage (Implicit)

See Also

VkPhysicalDeviceMemoryProperties, VkStructureType, vkGetPhysicalDeviceMemoryProperties2, vkGetPhysicalDeviceMemoryProperties2KHR

Constructors

PhysicalDeviceMemoryProperties2

Fields

next :: Maybe SomeVkStruct

memoryProperties :: PhysicalDeviceMemoryProperties

withCStructPhysicalDeviceProperties2 :: PhysicalDeviceProperties2 -> (VkPhysicalDeviceProperties2 -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceProperties2 and marshal a PhysicalDeviceProperties2 into it. The VkPhysicalDeviceProperties2 is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceProperties2 :: VkPhysicalDeviceProperties2 -> IO PhysicalDeviceProperties2

A function to read a VkPhysicalDeviceProperties2 and all additional structures in the pointer chain into a PhysicalDeviceProperties2.

data PhysicalDeviceProperties2

VkPhysicalDeviceProperties2 - Structure specifying physical device properties

Description

The pNext chain of this structure is used to extend the structure with properties defined by extensions.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PROPERTIES_2

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkPhysicalDeviceBlendOperationAdvancedPropertiesEXT, VkPhysicalDeviceConservativeRasterizationPropertiesEXT, VkPhysicalDeviceCooperativeMatrixPropertiesNV, VkPhysicalDeviceDepthStencilResolvePropertiesKHR, VkPhysicalDeviceDescriptorIndexingPropertiesEXT, VkPhysicalDeviceDiscardRectanglePropertiesEXT, VkPhysicalDeviceDriverPropertiesKHR, VkPhysicalDeviceExternalMemoryHostPropertiesEXT, VkPhysicalDeviceFloatControlsPropertiesKHR, VkPhysicalDeviceFragmentDensityMapPropertiesEXT, VkPhysicalDeviceIDProperties, VkPhysicalDeviceInlineUniformBlockPropertiesEXT, VkPhysicalDeviceMaintenance3Properties, VkPhysicalDeviceMeshShaderPropertiesNV, VkPhysicalDeviceMultiviewPerViewAttributesPropertiesNVX, VkPhysicalDeviceMultiviewProperties, VkPhysicalDevicePCIBusInfoPropertiesEXT, VkPhysicalDevicePointClippingProperties, VkPhysicalDeviceProtectedMemoryProperties, VkPhysicalDevicePushDescriptorPropertiesKHR, VkPhysicalDeviceRayTracingPropertiesNV, VkPhysicalDeviceSampleLocationsPropertiesEXT, VkPhysicalDeviceSamplerFilterMinmaxPropertiesEXT, VkPhysicalDeviceShaderCorePropertiesAMD, VkPhysicalDeviceShadingRateImagePropertiesNV, VkPhysicalDeviceSubgroupProperties, VkPhysicalDeviceTransformFeedbackPropertiesEXT, or VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT

Each sType member in the pNext chain must be unique

See Also

VkPhysicalDeviceProperties, VkStructureType, vkGetPhysicalDeviceProperties2, vkGetPhysicalDeviceProperties2KHR

Constructors

PhysicalDeviceProperties2

Fields

next :: Maybe SomeVkStruct

properties :: PhysicalDeviceProperties

withCStructPhysicalDeviceSparseImageFormatInfo2 :: PhysicalDeviceSparseImageFormatInfo2 -> (VkPhysicalDeviceSparseImageFormatInfo2 -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceSparseImageFormatInfo2 and marshal a PhysicalDeviceSparseImageFormatInfo2 into it. The VkPhysicalDeviceSparseImageFormatInfo2 is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceSparseImageFormatInfo2 :: VkPhysicalDeviceSparseImageFormatInfo2 -> IO PhysicalDeviceSparseImageFormatInfo2

A function to read a VkPhysicalDeviceSparseImageFormatInfo2 and all additional structures in the pointer chain into a PhysicalDeviceSparseImageFormatInfo2.

data PhysicalDeviceSparseImageFormatInfo2

VkPhysicalDeviceSparseImageFormatInfo2 - Structure specifying sparse image format inputs

Valid Usage (Implicit)

See Also

VkFormat, VkImageTiling, VkImageType, VkImageUsageFlags, VkSampleCountFlagBits, VkStructureType, vkGetPhysicalDeviceSparseImageFormatProperties2, vkGetPhysicalDeviceSparseImageFormatProperties2KHR

Constructors

PhysicalDeviceSparseImageFormatInfo2

Fields

next :: Maybe SomeVkStruct

format :: Format

type' :: ImageType

samples :: SampleCountFlagBits

usage :: ImageUsageFlags

tiling :: ImageTiling

withCStructQueueFamilyProperties2 :: QueueFamilyProperties2 -> (VkQueueFamilyProperties2 -> IO a) -> IO a

A function to temporarily allocate memory for a VkQueueFamilyProperties2 and marshal a QueueFamilyProperties2 into it. The VkQueueFamilyProperties2 is only valid inside the provided computation and must not be returned out of it.

fromCStructQueueFamilyProperties2 :: VkQueueFamilyProperties2 -> IO QueueFamilyProperties2

A function to read a VkQueueFamilyProperties2 and all additional structures in the pointer chain into a QueueFamilyProperties2.

data QueueFamilyProperties2

VkQueueFamilyProperties2 - Structure providing information about a queue family

Valid Usage (Implicit)

See Also

VkQueueFamilyProperties, VkStructureType, vkGetPhysicalDeviceQueueFamilyProperties2, vkGetPhysicalDeviceQueueFamilyProperties2KHR

Constructors

QueueFamilyProperties2

Fields

next :: Maybe SomeVkStruct

queueFamilyProperties :: QueueFamilyProperties

withCStructSparseImageFormatProperties2 :: SparseImageFormatProperties2 -> (VkSparseImageFormatProperties2 -> IO a) -> IO a

A function to temporarily allocate memory for a VkSparseImageFormatProperties2 and marshal a SparseImageFormatProperties2 into it. The VkSparseImageFormatProperties2 is only valid inside the provided computation and must not be returned out of it.

fromCStructSparseImageFormatProperties2 :: VkSparseImageFormatProperties2 -> IO SparseImageFormatProperties2

A function to read a VkSparseImageFormatProperties2 and all additional structures in the pointer chain into a SparseImageFormatProperties2.

data SparseImageFormatProperties2

VkSparseImageFormatProperties2 - Structure specifying sparse image format properties

Valid Usage (Implicit)

See Also

VkSparseImageFormatProperties, VkStructureType, vkGetPhysicalDeviceSparseImageFormatProperties2, vkGetPhysicalDeviceSparseImageFormatProperties2KHR

Constructors

SparseImageFormatProperties2

Fields

next :: Maybe SomeVkStruct

properties :: SparseImageFormatProperties

getPhysicalDeviceFeatures2 :: PhysicalDevice -> IO PhysicalDeviceFeatures2

vkGetPhysicalDeviceFeatures2 - Reports capabilities of a physical device

Parameters

physicalDevice is the physical device from which to query the supported features.

pFeatures is a pointer to a VkPhysicalDeviceFeatures2 structure in which the physical device features are returned.

Description

Each structure in pFeatures and its pNext chain contain members corresponding to fine-grained features. vkGetPhysicalDeviceFeatures2 writes each member to a boolean value indicating whether that feature is supported.

Valid Usage (Implicit)

See Also

VkPhysicalDevice, VkPhysicalDeviceFeatures2

getPhysicalDeviceFormatProperties2 :: PhysicalDevice -> Format -> IO FormatProperties2

vkGetPhysicalDeviceFormatProperties2 - Lists physical device’s format capabilities

Parameters

physicalDevice is the physical device from which to query the format properties.

format is the format whose properties are queried.

pFormatProperties is a pointer to a VkFormatProperties2 structure in which physical device properties for format are returned.

Description

vkGetPhysicalDeviceFormatProperties2 behaves similarly to vkGetPhysicalDeviceFormatProperties, with the ability to return extended information in a pNext chain of output structures.

Valid Usage (Implicit)

See Also

VkFormat, VkFormatProperties2, VkPhysicalDevice

getPhysicalDeviceImageFormatProperties2 :: PhysicalDevice -> PhysicalDeviceImageFormatInfo2 -> IO ImageFormatProperties2

vkGetPhysicalDeviceImageFormatProperties2 - Lists physical device’s image format capabilities

Parameters

physicalDevice is the physical device from which to query the image capabilities.

pImageFormatInfo points to an instance of the VkPhysicalDeviceImageFormatInfo2 structure, describing the parameters that would be consumed by vkCreateImage.

pImageFormatProperties points to an instance of the VkImageFormatProperties2 structure in which capabilities are returned.

Description

vkGetPhysicalDeviceImageFormatProperties2 behaves similarly to vkGetPhysicalDeviceImageFormatProperties, with the ability to return extended information in a pNext chain of output structures.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_FORMAT_NOT_SUPPORTED

See Also

VkImageFormatProperties2, VkPhysicalDevice, VkPhysicalDeviceImageFormatInfo2

getPhysicalDeviceMemoryProperties2 :: PhysicalDevice -> IO PhysicalDeviceMemoryProperties2

vkGetPhysicalDeviceMemoryProperties2 - Reports memory information for the specified physical device

Parameters

physicalDevice is the handle to the device to query.

pMemoryProperties points to an instance of VkPhysicalDeviceMemoryProperties2 structure in which the properties are returned.

Description

vkGetPhysicalDeviceMemoryProperties2 behaves similarly to vkGetPhysicalDeviceMemoryProperties, with the ability to return extended information in a pNext chain of output structures.

Valid Usage (Implicit)

See Also

VkPhysicalDevice, VkPhysicalDeviceMemoryProperties2

getPhysicalDeviceProperties2 :: PhysicalDevice -> IO PhysicalDeviceProperties2

vkGetPhysicalDeviceProperties2 - Returns properties of a physical device

Parameters

physicalDevice is the handle to the physical device whose properties will be queried.

pProperties points to an instance of the VkPhysicalDeviceProperties2 structure, that will be filled with returned information.

Description

Each structure in pProperties and its pNext chain contain members corresponding to properties or implementation-dependent limits. vkGetPhysicalDeviceProperties2 writes each member to a value indicating the value of that property or limit.

Valid Usage (Implicit)

See Also

VkPhysicalDevice, VkPhysicalDeviceProperties2

getNumPhysicalDeviceQueueFamilyProperties2 :: PhysicalDevice -> IO Word32

vkGetPhysicalDeviceQueueFamilyProperties2 - Reports properties of the queues of the specified physical device

Parameters

physicalDevice is the handle to the physical device whose properties will be queried.

pQueueFamilyPropertyCount is a pointer to an integer related to the number of queue families available or queried, as described in vkGetPhysicalDeviceQueueFamilyProperties.

pQueueFamilyProperties is either NULL or a pointer to an array of VkQueueFamilyProperties2 structures.

Description

vkGetPhysicalDeviceQueueFamilyProperties2 behaves similarly to vkGetPhysicalDeviceQueueFamilyProperties, with the ability to return extended information in a pNext chain of output structures.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pQueueFamilyPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pQueueFamilyPropertyCount is not 0, and pQueueFamilyProperties is not NULL, pQueueFamilyProperties must be a valid pointer to an array of pQueueFamilyPropertyCount VkQueueFamilyProperties2 structures

See Also

VkPhysicalDevice, VkQueueFamilyProperties2

getPhysicalDeviceQueueFamilyProperties2 :: PhysicalDevice -> Word32 -> IO (Vector QueueFamilyProperties2)

vkGetPhysicalDeviceQueueFamilyProperties2 - Reports properties of the queues of the specified physical device

Parameters

physicalDevice is the handle to the physical device whose properties will be queried.

pQueueFamilyPropertyCount is a pointer to an integer related to the number of queue families available or queried, as described in vkGetPhysicalDeviceQueueFamilyProperties.

pQueueFamilyProperties is either NULL or a pointer to an array of VkQueueFamilyProperties2 structures.

Description

vkGetPhysicalDeviceQueueFamilyProperties2 behaves similarly to vkGetPhysicalDeviceQueueFamilyProperties, with the ability to return extended information in a pNext chain of output structures.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pQueueFamilyPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pQueueFamilyPropertyCount is not 0, and pQueueFamilyProperties is not NULL, pQueueFamilyProperties must be a valid pointer to an array of pQueueFamilyPropertyCount VkQueueFamilyProperties2 structures

See Also

VkPhysicalDevice, VkQueueFamilyProperties2

getAllPhysicalDeviceQueueFamilyProperties2 :: PhysicalDevice -> IO (Vector QueueFamilyProperties2)

Returns all the values available from getPhysicalDeviceQueueFamilyProperties2.

getNumPhysicalDeviceSparseImageFormatProperties2 :: PhysicalDevice -> PhysicalDeviceSparseImageFormatInfo2 -> IO Word32

vkGetPhysicalDeviceSparseImageFormatProperties2 - Retrieve properties of an image format applied to sparse images

Parameters

physicalDevice is the physical device from which to query the sparse image capabilities.

pFormatInfo is a pointer to a structure of type VkPhysicalDeviceSparseImageFormatInfo2 containing input parameters to the command.

pPropertyCount is a pointer to an integer related to the number of sparse format properties available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkSparseImageFormatProperties2 structures.

Description

vkGetPhysicalDeviceSparseImageFormatProperties2 behaves identically to vkGetPhysicalDeviceSparseImageFormatProperties, with the ability to return extended information by adding extension structures to the pNext chain of its pProperties parameter.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pFormatInfo must be a valid pointer to a valid VkPhysicalDeviceSparseImageFormatInfo2 structure

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkSparseImageFormatProperties2 structures

See Also

VkPhysicalDevice, VkPhysicalDeviceSparseImageFormatInfo2, VkSparseImageFormatProperties2

getPhysicalDeviceSparseImageFormatProperties2 :: PhysicalDevice -> PhysicalDeviceSparseImageFormatInfo2 -> Word32 -> IO (Vector SparseImageFormatProperties2)

vkGetPhysicalDeviceSparseImageFormatProperties2 - Retrieve properties of an image format applied to sparse images

Parameters

physicalDevice is the physical device from which to query the sparse image capabilities.

pFormatInfo is a pointer to a structure of type VkPhysicalDeviceSparseImageFormatInfo2 containing input parameters to the command.

pPropertyCount is a pointer to an integer related to the number of sparse format properties available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkSparseImageFormatProperties2 structures.

Description

vkGetPhysicalDeviceSparseImageFormatProperties2 behaves identically to vkGetPhysicalDeviceSparseImageFormatProperties, with the ability to return extended information by adding extension structures to the pNext chain of its pProperties parameter.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pFormatInfo must be a valid pointer to a valid VkPhysicalDeviceSparseImageFormatInfo2 structure

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkSparseImageFormatProperties2 structures

See Also

VkPhysicalDevice, VkPhysicalDeviceSparseImageFormatInfo2, VkSparseImageFormatProperties2

getAllPhysicalDeviceSparseImageFormatProperties2 :: PhysicalDevice -> PhysicalDeviceSparseImageFormatInfo2 -> IO (Vector SparseImageFormatProperties2)

Returns all the values available from getPhysicalDeviceSparseImageFormatProperties2.

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_FEATURES_2 :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_PROPERTIES_2 :: a ~ StructureType => a

pattern STRUCTURE_TYPE_FORMAT_PROPERTIES_2 :: a ~ StructureType => a

pattern STRUCTURE_TYPE_IMAGE_FORMAT_PROPERTIES_2 :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGE_FORMAT_INFO_2 :: a ~ StructureType => a

pattern STRUCTURE_TYPE_QUEUE_FAMILY_PROPERTIES_2 :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_MEMORY_PROPERTIES_2 :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SPARSE_IMAGE_FORMAT_PROPERTIES_2 :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SPARSE_IMAGE_FORMAT_INFO_2 :: a ~ StructureType => a