Graphics.Vulkan.Core11.Promoted_from_VK_KHR_maintenance2

Documentation

withCStructImageViewUsageCreateInfo :: ImageViewUsageCreateInfo -> (VkImageViewUsageCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkImageViewUsageCreateInfo and marshal a ImageViewUsageCreateInfo into it. The VkImageViewUsageCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructImageViewUsageCreateInfo :: VkImageViewUsageCreateInfo -> IO ImageViewUsageCreateInfo

A function to read a VkImageViewUsageCreateInfo and all additional structures in the pointer chain into a ImageViewUsageCreateInfo.

data ImageViewUsageCreateInfo

VkImageViewUsageCreateInfo - Specify the intended usage of an image view

Description

When this structure is chained to VkImageViewCreateInfo the usage field overrides the implicit usage parameter inherited from image creation time and its value is used instead for the purposes of determining the valid usage conditions of VkImageViewCreateInfo.

Valid Usage (Implicit)

See Also

VkImageUsageFlags, VkStructureType

Constructors

ImageViewUsageCreateInfo

Fields

next :: Maybe SomeVkStruct

usage :: ImageUsageFlags

withCStructInputAttachmentAspectReference :: InputAttachmentAspectReference -> (VkInputAttachmentAspectReference -> IO a) -> IO a

A function to temporarily allocate memory for a VkInputAttachmentAspectReference and marshal a InputAttachmentAspectReference into it. The VkInputAttachmentAspectReference is only valid inside the provided computation and must not be returned out of it.

fromCStructInputAttachmentAspectReference :: VkInputAttachmentAspectReference -> IO InputAttachmentAspectReference

A function to read a VkInputAttachmentAspectReference and all additional structures in the pointer chain into a InputAttachmentAspectReference.

data InputAttachmentAspectReference

VkInputAttachmentAspectReference - Structure specifying a subpass/input attachment pair and an aspect mask that can be read.

Valid Usage (Implicit)

See Also

VkImageAspectFlags, VkRenderPassInputAttachmentAspectCreateInfo

Constructors

InputAttachmentAspectReference

Fields

subpass :: Word32

inputAttachmentIndex :: Word32

aspectMask :: ImageAspectFlags

withCStructPhysicalDevicePointClippingProperties :: PhysicalDevicePointClippingProperties -> (VkPhysicalDevicePointClippingProperties -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDevicePointClippingProperties and marshal a PhysicalDevicePointClippingProperties into it. The VkPhysicalDevicePointClippingProperties is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDevicePointClippingProperties :: VkPhysicalDevicePointClippingProperties -> IO PhysicalDevicePointClippingProperties

A function to read a VkPhysicalDevicePointClippingProperties and all additional structures in the pointer chain into a PhysicalDevicePointClippingProperties.

data PhysicalDevicePointClippingProperties

VkPhysicalDevicePointClippingProperties - Structure describing the point clipping behavior supported by an implementation

Members

The members of the VkPhysicalDevicePointClippingProperties structure describe the following implementation-dependent limit:

Description

If the VkPhysicalDevicePointClippingProperties structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkPointClippingBehavior, VkStructureType

Constructors

PhysicalDevicePointClippingProperties

Fields

next :: Maybe SomeVkStruct

pointClippingBehavior :: PointClippingBehavior

withCStructPipelineTessellationDomainOriginStateCreateInfo :: PipelineTessellationDomainOriginStateCreateInfo -> (VkPipelineTessellationDomainOriginStateCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkPipelineTessellationDomainOriginStateCreateInfo and marshal a PipelineTessellationDomainOriginStateCreateInfo into it. The VkPipelineTessellationDomainOriginStateCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructPipelineTessellationDomainOriginStateCreateInfo :: VkPipelineTessellationDomainOriginStateCreateInfo -> IO PipelineTessellationDomainOriginStateCreateInfo

A function to read a VkPipelineTessellationDomainOriginStateCreateInfo and all additional structures in the pointer chain into a PipelineTessellationDomainOriginStateCreateInfo.

data PipelineTessellationDomainOriginStateCreateInfo

VkPipelineTessellationDomainOriginStateCreateInfo - Structure specifying the orientation of the tessellation domain

Description

If the VkPipelineTessellationDomainOriginStateCreateInfo structure is included in the pNext chain of VkPipelineTessellationStateCreateInfo, it controls the origin of the tessellation domain. If this structure is not present, it is as if domainOrigin were VK_TESSELLATION_DOMAIN_ORIGIN_UPPER_LEFT.

Valid Usage (Implicit)

See Also

VkStructureType, VkTessellationDomainOrigin

Constructors

PipelineTessellationDomainOriginStateCreateInfo

Fields

next :: Maybe SomeVkStruct

domainOrigin :: TessellationDomainOrigin

type PointClippingBehavior = VkPointClippingBehavior

VkPointClippingBehavior - Enum specifying the point clipping behavior

See Also

VkPhysicalDevicePointClippingProperties

pattern POINT_CLIPPING_BEHAVIOR_ALL_CLIP_PLANES :: a ~ PointClippingBehavior => a

VK_POINT_CLIPPING_BEHAVIOR_ALL_CLIP_PLANES specifies that the primitive is discarded if the vertex lies outside any clip plane, including the planes bounding the view volume.

pattern POINT_CLIPPING_BEHAVIOR_USER_CLIP_PLANES_ONLY :: a ~ PointClippingBehavior => a

VK_POINT_CLIPPING_BEHAVIOR_USER_CLIP_PLANES_ONLY specifies that the primitive is discarded only if the vertex lies outside any user clip plane.

type PointClippingBehaviorKHR = PointClippingBehavior

withCStructRenderPassInputAttachmentAspectCreateInfo :: RenderPassInputAttachmentAspectCreateInfo -> (VkRenderPassInputAttachmentAspectCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkRenderPassInputAttachmentAspectCreateInfo and marshal a RenderPassInputAttachmentAspectCreateInfo into it. The VkRenderPassInputAttachmentAspectCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructRenderPassInputAttachmentAspectCreateInfo :: VkRenderPassInputAttachmentAspectCreateInfo -> IO RenderPassInputAttachmentAspectCreateInfo

A function to read a VkRenderPassInputAttachmentAspectCreateInfo and all additional structures in the pointer chain into a RenderPassInputAttachmentAspectCreateInfo.

data RenderPassInputAttachmentAspectCreateInfo

VkRenderPassInputAttachmentAspectCreateInfo - Structure specifying, for a given subpass/input attachment pair, which aspect can be read.

Valid Usage (Implicit)

See Also

VkInputAttachmentAspectReference, VkStructureType

Constructors

RenderPassInputAttachmentAspectCreateInfo

Fields

next :: Maybe SomeVkStruct

aspectReferences :: Vector InputAttachmentAspectReference

type TessellationDomainOrigin = VkTessellationDomainOrigin

VkTessellationDomainOrigin - Enum describing tessellation domain origin

Description

This enum affects how the VertexOrderCw and VertexOrderCcw tessellation execution modes are interpreted, since the winding is defined relative to the orientation of the domain.

See Also

VkPipelineTessellationDomainOriginStateCreateInfo

pattern TESSELLATION_DOMAIN_ORIGIN_UPPER_LEFT :: a ~ TessellationDomainOrigin => a

VK_TESSELLATION_DOMAIN_ORIGIN_UPPER_LEFT specifies that the origin of the domain space is in the upper left corner, as shown in figure https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#img-tessellation-topology-ul.

pattern TESSELLATION_DOMAIN_ORIGIN_LOWER_LEFT :: a ~ TessellationDomainOrigin => a

VK_TESSELLATION_DOMAIN_ORIGIN_LOWER_LEFT specifies that the origin of the domain space is in the lower left corner, as shown in figure https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#img-tessellation-topology-ll.

type TessellationDomainOriginKHR = TessellationDomainOrigin

pattern IMAGE_CREATE_BLOCK_TEXEL_VIEW_COMPATIBLE_BIT :: a ~ ImageCreateFlagBits => a

pattern IMAGE_CREATE_EXTENDED_USAGE_BIT :: a ~ ImageCreateFlagBits => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_POINT_CLIPPING_PROPERTIES :: a ~ StructureType => a

pattern STRUCTURE_TYPE_RENDER_PASS_INPUT_ATTACHMENT_ASPECT_CREATE_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_IMAGE_VIEW_USAGE_CREATE_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PIPELINE_TESSELLATION_DOMAIN_ORIGIN_STATE_CREATE_INFO :: a ~ StructureType => a

pattern IMAGE_LAYOUT_DEPTH_READ_ONLY_STENCIL_ATTACHMENT_OPTIMAL :: a ~ ImageLayout => a

pattern IMAGE_LAYOUT_DEPTH_ATTACHMENT_STENCIL_READ_ONLY_OPTIMAL :: a ~ ImageLayout => a