Graphics.Vulkan.Core11.Promoted_from_VK_KHR_maintenance3

Documentation

withCStructDescriptorSetLayoutSupport :: DescriptorSetLayoutSupport -> (VkDescriptorSetLayoutSupport -> IO a) -> IO a

A function to temporarily allocate memory for a VkDescriptorSetLayoutSupport and marshal a DescriptorSetLayoutSupport into it. The VkDescriptorSetLayoutSupport is only valid inside the provided computation and must not be returned out of it.

fromCStructDescriptorSetLayoutSupport :: VkDescriptorSetLayoutSupport -> IO DescriptorSetLayoutSupport

A function to read a VkDescriptorSetLayoutSupport and all additional structures in the pointer chain into a DescriptorSetLayoutSupport.

data DescriptorSetLayoutSupport

VkDescriptorSetLayoutSupport - Structure returning information about whether a descriptor set layout can be supported

Description

supported is set to VK_TRUE if the descriptor set can be created, or else is set to VK_FALSE.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType, vkGetDescriptorSetLayoutSupport, vkGetDescriptorSetLayoutSupportKHR

Constructors

DescriptorSetLayoutSupport

Fields

next :: Maybe SomeVkStruct

supported :: Bool

withCStructPhysicalDeviceMaintenance3Properties :: PhysicalDeviceMaintenance3Properties -> (VkPhysicalDeviceMaintenance3Properties -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceMaintenance3Properties and marshal a PhysicalDeviceMaintenance3Properties into it. The VkPhysicalDeviceMaintenance3Properties is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceMaintenance3Properties :: VkPhysicalDeviceMaintenance3Properties -> IO PhysicalDeviceMaintenance3Properties

A function to read a VkPhysicalDeviceMaintenance3Properties and all additional structures in the pointer chain into a PhysicalDeviceMaintenance3Properties.

data PhysicalDeviceMaintenance3Properties

VkPhysicalDeviceMaintenance3Properties - Structure describing descriptor set properties

Members

The members of the VkPhysicalDeviceMaintenance3Properties structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceMaintenance3Properties structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkDeviceSize, VkStructureType

Constructors

PhysicalDeviceMaintenance3Properties

Fields

next :: Maybe SomeVkStruct

maxPerSetDescriptors :: Word32

maxMemoryAllocationSize :: DeviceSize

getDescriptorSetLayoutSupport :: Device -> DescriptorSetLayoutCreateInfo -> IO DescriptorSetLayoutSupport

vkGetDescriptorSetLayoutSupport - Query whether a descriptor set layout can be created

Parameters

device is the logical device that would create the descriptor set layout.

pCreateInfo is a pointer to an instance of the VkDescriptorSetLayoutCreateInfo structure specifying the state of the descriptor set layout object.

pSupport points to a VkDescriptorSetLayoutSupport structure in which information about support for the descriptor set layout object is returned.

Description

Some implementations have limitations on what fits in a descriptor set which are not easily expressible in terms of existing limits like maxDescriptorSet*, for example if all descriptor types share a limited space in memory but each descriptor is a different size or alignment. This command returns information about whether a descriptor set satisfies this limit. If the descriptor set layout satisfies the VkPhysicalDeviceMaintenance3Properties::maxPerSetDescriptors limit, this command is guaranteed to return VK_TRUE in VkDescriptorSetLayoutSupport::supported. If the descriptor set layout exceeds the VkPhysicalDeviceMaintenance3Properties::maxPerSetDescriptors limit, whether the descriptor set layout is supported is implementation-dependent and may depend on whether the descriptor sizes and alignments cause the layout to exceed an internal limit.

This command does not consider other limits such as maxPerStageDescriptor*, and so a descriptor set layout that is supported according to this command must still satisfy the pipeline layout limits such as maxPerStageDescriptor* in order to be used in a pipeline layout.

Note

This is a VkDevice query rather than VkPhysicalDevice because the answer may depend on enabled features.

Valid Usage (Implicit)

See Also

VkDescriptorSetLayoutCreateInfo, VkDescriptorSetLayoutSupport, VkDevice

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_MAINTENANCE_3_PROPERTIES :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_SUPPORT :: a ~ StructureType => a