Graphics.Vulkan.Core11.Promoted_from_VK_KHR_multiview

Documentation

withCStructPhysicalDeviceMultiviewFeatures :: PhysicalDeviceMultiviewFeatures -> (VkPhysicalDeviceMultiviewFeatures -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceMultiviewFeatures and marshal a PhysicalDeviceMultiviewFeatures into it. The VkPhysicalDeviceMultiviewFeatures is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceMultiviewFeatures :: VkPhysicalDeviceMultiviewFeatures -> IO PhysicalDeviceMultiviewFeatures

A function to read a VkPhysicalDeviceMultiviewFeatures and all additional structures in the pointer chain into a PhysicalDeviceMultiviewFeatures.

data PhysicalDeviceMultiviewFeatures

VkPhysicalDeviceMultiviewFeatures - Structure describing multiview features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceMultiviewFeatures structure describe the following features:

Description

multiview specifies whether the implementation supports multiview rendering within a render pass. If this feature is not enabled, the view mask of each subpass must always be zero.

multiviewGeometryShader specifies whether the implementation supports multiview rendering within a render pass, with geometry shaders. If this feature is not enabled, then a pipeline compiled against a subpass with a non-zero view mask must not include a geometry shader.

multiviewTessellationShader specifies whether the implementation supports multiview rendering within a render pass, with tessellation shaders. If this feature is not enabled, then a pipeline compiled against a subpass with a non-zero view mask must not include any tessellation shaders.

If the VkPhysicalDeviceMultiviewFeatures structure is included in the pNext chain of VkPhysicalDeviceFeatures2, it is filled with values indicating whether each feature is supported. VkPhysicalDeviceMultiviewFeatures can also be used in the pNext chain of VkDeviceCreateInfo to enable the features.

Valid Usage

If multiviewGeometryShader is enabled then multiview must also be enabled.

If multiviewTessellationShader is enabled then multiview must also be enabled.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MULTIVIEW_FEATURES

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceMultiviewFeatures

Fields

next :: Maybe SomeVkStruct

multiview :: Bool

multiviewGeometryShader :: Bool

multiviewTessellationShader :: Bool

withCStructPhysicalDeviceMultiviewProperties :: PhysicalDeviceMultiviewProperties -> (VkPhysicalDeviceMultiviewProperties -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceMultiviewProperties and marshal a PhysicalDeviceMultiviewProperties into it. The VkPhysicalDeviceMultiviewProperties is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceMultiviewProperties :: VkPhysicalDeviceMultiviewProperties -> IO PhysicalDeviceMultiviewProperties

A function to read a VkPhysicalDeviceMultiviewProperties and all additional structures in the pointer chain into a PhysicalDeviceMultiviewProperties.

data PhysicalDeviceMultiviewProperties

VkPhysicalDeviceMultiviewProperties - Structure describing multiview limits that can be supported by an implementation

Members

The members of the VkPhysicalDeviceMultiviewProperties structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceMultiviewProperties structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

PhysicalDeviceMultiviewProperties

Fields

next :: Maybe SomeVkStruct

maxMultiviewViewCount :: Word32

maxMultiviewInstanceIndex :: Word32

withCStructRenderPassMultiviewCreateInfo :: RenderPassMultiviewCreateInfo -> (VkRenderPassMultiviewCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkRenderPassMultiviewCreateInfo and marshal a RenderPassMultiviewCreateInfo into it. The VkRenderPassMultiviewCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructRenderPassMultiviewCreateInfo :: VkRenderPassMultiviewCreateInfo -> IO RenderPassMultiviewCreateInfo

A function to read a VkRenderPassMultiviewCreateInfo and all additional structures in the pointer chain into a RenderPassMultiviewCreateInfo.

data RenderPassMultiviewCreateInfo

VkRenderPassMultiviewCreateInfo - Structure containing multiview info for all subpasses

Description

When a subpass uses a non-zero view mask, multiview functionality is considered to be enabled. Multiview is all-or-nothing for a render pass - that is, either all subpasses must have a non-zero view mask (though some subpasses may have only one view) or all must be zero. Multiview causes all drawing and clear commands in the subpass to behave as if they were broadcast to each view, where a view is represented by one layer of the framebuffer attachments. All draws and clears are broadcast to each view index whose bit is set in the view mask. The view index is provided in the ViewIndex shader input variable, and color, depth/stencil, and input attachments all read/write the layer of the framebuffer corresponding to the view index.

If the view mask is zero for all subpasses, multiview is considered to be disabled and all drawing commands execute normally, without this additional broadcasting.

Some implementations may not support multiview in conjunction with geometry shaders or tessellation shaders.

When multiview is enabled, the VK_DEPENDENCY_VIEW_LOCAL_BIT bit in a dependency can be used to express a view-local dependency, meaning that each view in the destination subpass depends on a single view in the source subpass. Unlike pipeline barriers, a subpass dependency can potentially have a different view mask in the source subpass and the destination subpass. If the dependency is view-local, then each view (dstView) in the destination subpass depends on the view dstView + pViewOffsets[dependency] in the source subpass. If there is not such a view in the source subpass, then this dependency does not affect that view in the destination subpass. If the dependency is not view-local, then all views in the destination subpass depend on all views in the source subpass, and the view offset is ignored. A non-zero view offset is not allowed in a self-dependency.

The elements of pCorrelationMasks are a set of masks of views indicating that views in the same mask may exhibit spatial coherency between the views, making it more efficient to render them concurrently. Correlation masks must not have a functional effect on the results of the multiview rendering.

When multiview is enabled, at the beginning of each subpass all non-render pass state is undefined. In particular, each time vkCmdBeginRenderPass or vkCmdNextSubpass is called the graphics pipeline must be bound, any relevant descriptor sets or vertex/index buffers must be bound, and any relevant dynamic state or push constants must be set before they are used.

Valid Usage

Each view index must not be set in more than one element of pCorrelationMasks

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_RENDER_PASS_MULTIVIEW_CREATE_INFO

If subpassCount is not 0, pViewMasks must be a valid pointer to an array of subpassCount uint32_t values

If dependencyCount is not 0, pViewOffsets must be a valid pointer to an array of dependencyCount int32_t values

If correlationMaskCount is not 0, pCorrelationMasks must be a valid pointer to an array of correlationMaskCount uint32_t values

See Also

VkStructureType

Constructors

RenderPassMultiviewCreateInfo

Fields

next :: Maybe SomeVkStruct

viewMasks :: Vector Word32

viewOffsets :: Vector Int32

correlationMasks :: Vector Word32

pattern STRUCTURE_TYPE_RENDER_PASS_MULTIVIEW_CREATE_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_MULTIVIEW_FEATURES :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_MULTIVIEW_PROPERTIES :: a ~ StructureType => a

pattern DEPENDENCY_VIEW_LOCAL_BIT :: a ~ DependencyFlagBits => a