Graphics.Vulkan.Core11.Promoted_from_VK_KHR_sampler_ycbcr_conversion

Documentation

withCStructBindImagePlaneMemoryInfo :: BindImagePlaneMemoryInfo -> (VkBindImagePlaneMemoryInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkBindImagePlaneMemoryInfo and marshal a BindImagePlaneMemoryInfo into it. The VkBindImagePlaneMemoryInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructBindImagePlaneMemoryInfo :: VkBindImagePlaneMemoryInfo -> IO BindImagePlaneMemoryInfo

A function to read a VkBindImagePlaneMemoryInfo and all additional structures in the pointer chain into a BindImagePlaneMemoryInfo.

data BindImagePlaneMemoryInfo

VkBindImagePlaneMemoryInfo - Structure specifying how to bind an image plane to memory

Valid Usage

If the image’s tiling is VK_IMAGE_TILING_LINEAR or VK_IMAGE_TILING_OPTIMAL, then planeAspect must be a single valid format plane for the image. (That is, planeAspect must be VK_IMAGE_ASPECT_PLANE_0_BIT or VK_IMAGE_ASPECT_PLANE_1_BIT for “_2PLANE” formats and planeAspect must be VK_IMAGE_ASPECT_PLANE_0_BIT, VK_IMAGE_ASPECT_PLANE_1_BIT, or VK_IMAGE_ASPECT_PLANE_2_BIT for “_3PLANE” formats.)

A single call to vkBindImageMemory2 must bind all or none of the planes of an image (i.e. bindings to all planes of an image must be made in a single vkBindImageMemory2 call), as separate bindings

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_BIND_IMAGE_PLANE_MEMORY_INFO

planeAspect must be a valid VkImageAspectFlagBits value

See Also

VkImageAspectFlagBits, VkStructureType

Constructors

BindImagePlaneMemoryInfo

Fields

next :: Maybe SomeVkStruct

planeAspect :: ImageAspectFlagBits

type ChromaLocation = VkChromaLocation

VkChromaLocation - Position of downsampled chroma samples

See Also

VkAndroidHardwareBufferFormatPropertiesANDROID, VkSamplerYcbcrConversionCreateInfo

pattern CHROMA_LOCATION_COSITED_EVEN :: a ~ ChromaLocation => a

VK_CHROMA_LOCATION_COSITED_EVEN specifies that downsampled chroma samples are aligned with luma samples with even coordinates.

pattern CHROMA_LOCATION_MIDPOINT :: a ~ ChromaLocation => a

VK_CHROMA_LOCATION_MIDPOINT specifies that downsampled chroma samples are located half way between each even luma sample and the nearest higher odd luma sample.

type ChromaLocationKHR = ChromaLocation

withCStructImagePlaneMemoryRequirementsInfo :: ImagePlaneMemoryRequirementsInfo -> (VkImagePlaneMemoryRequirementsInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkImagePlaneMemoryRequirementsInfo and marshal a ImagePlaneMemoryRequirementsInfo into it. The VkImagePlaneMemoryRequirementsInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructImagePlaneMemoryRequirementsInfo :: VkImagePlaneMemoryRequirementsInfo -> IO ImagePlaneMemoryRequirementsInfo

A function to read a VkImagePlaneMemoryRequirementsInfo and all additional structures in the pointer chain into a ImagePlaneMemoryRequirementsInfo.

data ImagePlaneMemoryRequirementsInfo

VkImagePlaneMemoryRequirementsInfo - Structure specifying image plane for memory requirements

Valid Usage

If the image’s tiling is VK_IMAGE_TILING_LINEAR or VK_IMAGE_TILING_OPTIMAL, then planeAspect must be a single valid format plane for the image. (That is, for a two-plane image planeAspect must be VK_IMAGE_ASPECT_PLANE_0_BIT or VK_IMAGE_ASPECT_PLANE_1_BIT, and for a three-plane image planeAspect must be VK_IMAGE_ASPECT_PLANE_0_BIT, VK_IMAGE_ASPECT_PLANE_1_BIT or VK_IMAGE_ASPECT_PLANE_2_BIT).

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMAGE_PLANE_MEMORY_REQUIREMENTS_INFO

planeAspect must be a valid VkImageAspectFlagBits value

See Also

VkImageAspectFlagBits, VkStructureType

Constructors

ImagePlaneMemoryRequirementsInfo

Fields

next :: Maybe SomeVkStruct

planeAspect :: ImageAspectFlagBits

withCStructPhysicalDeviceSamplerYcbcrConversionFeatures :: PhysicalDeviceSamplerYcbcrConversionFeatures -> (VkPhysicalDeviceSamplerYcbcrConversionFeatures -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceSamplerYcbcrConversionFeatures and marshal a PhysicalDeviceSamplerYcbcrConversionFeatures into it. The VkPhysicalDeviceSamplerYcbcrConversionFeatures is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceSamplerYcbcrConversionFeatures :: VkPhysicalDeviceSamplerYcbcrConversionFeatures -> IO PhysicalDeviceSamplerYcbcrConversionFeatures

A function to read a VkPhysicalDeviceSamplerYcbcrConversionFeatures and all additional structures in the pointer chain into a PhysicalDeviceSamplerYcbcrConversionFeatures.

data PhysicalDeviceSamplerYcbcrConversionFeatures

VkPhysicalDeviceSamplerYcbcrConversionFeatures - Structure describing Y’CbCr conversion features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceSamplerYcbcrConversionFeatures structure describe the following feature:

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceSamplerYcbcrConversionFeatures

Fields

next :: Maybe SomeVkStruct

samplerYcbcrConversion :: Bool

type SamplerYcbcrConversion = VkSamplerYcbcrConversion

VkSamplerYcbcrConversion - Opaque handle to a device-specific sampler Y’CBCR conversion description

See Also

VkSamplerYcbcrConversionInfo, vkCreateSamplerYcbcrConversion, vkCreateSamplerYcbcrConversionKHR, vkDestroySamplerYcbcrConversion, vkDestroySamplerYcbcrConversionKHR

withCStructSamplerYcbcrConversionCreateInfo :: SamplerYcbcrConversionCreateInfo -> (VkSamplerYcbcrConversionCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkSamplerYcbcrConversionCreateInfo and marshal a SamplerYcbcrConversionCreateInfo into it. The VkSamplerYcbcrConversionCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructSamplerYcbcrConversionCreateInfo :: VkSamplerYcbcrConversionCreateInfo -> IO SamplerYcbcrConversionCreateInfo

A function to read a VkSamplerYcbcrConversionCreateInfo and all additional structures in the pointer chain into a SamplerYcbcrConversionCreateInfo.

data SamplerYcbcrConversionCreateInfo

VkSamplerYcbcrConversionCreateInfo - Structure specifying the parameters of the newly created conversion

Description

Note

Setting forceExplicitReconstruction to VK_TRUE may have a performance penalty on implementations where explicit reconstruction is not the default mode of operation.

Sampler Y’CBCR conversion objects do not support /external format conversion without additional extensions defining external formats/.

Valid Usage

format must not be VK_FORMAT_UNDEFINED

format must support VK_FORMAT_FEATURE_MIDPOINT_CHROMA_SAMPLES_BIT or VK_FORMAT_FEATURE_COSITED_CHROMA_SAMPLES_BIT

If the format does not support VK_FORMAT_FEATURE_COSITED_CHROMA_SAMPLES_BIT, xChromaOffset and yChromaOffset must not be VK_CHROMA_LOCATION_COSITED_EVEN

If the format does not support VK_FORMAT_FEATURE_MIDPOINT_CHROMA_SAMPLES_BIT, xChromaOffset and yChromaOffset must not be VK_CHROMA_LOCATION_MIDPOINT

format must represent unsigned normalized values (i.e. the format must be a UNORM format)

If the format has a _422 or _420 suffix, then components.g must be VK_COMPONENT_SWIZZLE_IDENTITY

If the format has a _422 or _420 suffix, then components.a must be VK_COMPONENT_SWIZZLE_IDENTITY, VK_COMPONENT_SWIZZLE_ONE, or VK_COMPONENT_SWIZZLE_ZERO

If the format has a _422 or _420 suffix, then components.r must be VK_COMPONENT_SWIZZLE_IDENTITY or VK_COMPONENT_SWIZZLE_B

If the format has a _422 or _420 suffix, then components.b must be VK_COMPONENT_SWIZZLE_IDENTITY or VK_COMPONENT_SWIZZLE_R

If the format has a _422 or _420 suffix, and if either components.r or components.b is VK_COMPONENT_SWIZZLE_IDENTITY, both values must be VK_COMPONENT_SWIZZLE_IDENTITY

If ycbcrModel is not VK_SAMPLER_YCBCR_MODEL_CONVERSION_RGB_IDENTITY, then components.r, components.g, and components.b must correspond to channels of the format; that is, components.r, components.g, and components.b must not be VK_COMPONENT_SWIZZLE_ZERO or VK_COMPONENT_SWIZZLE_ONE, and must not correspond to a channel which contains zero or one as a consequence of conversion to RGBA

If the format does not support VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_FORCEABLE_BIT, forceExplicitReconstruction must be FALSE

If the format does not support VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_LINEAR_FILTER_BIT, chromaFilter must be VK_FILTER_NEAREST

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_CREATE_INFO

pNext must be NULL or a pointer to a valid instance of VkExternalFormatANDROID

format must be a valid VkFormat value

ycbcrModel must be a valid VkSamplerYcbcrModelConversion value

ycbcrRange must be a valid VkSamplerYcbcrRange value

components must be a valid VkComponentMapping structure

xChromaOffset must be a valid VkChromaLocation value

yChromaOffset must be a valid VkChromaLocation value

chromaFilter must be a valid VkFilter value

If chromaFilter is VK_FILTER_NEAREST, chroma samples are reconstructed to luma channel resolution using nearest-neighbour sampling. Otherwise, chroma samples are reconstructed using interpolation. More details can be found in the description of sampler Y’CBCR conversion in the Image Operations chapter.

See Also

VkBool32, VkChromaLocation, VkComponentMapping, VkFilter, VkFormat, VkSamplerYcbcrModelConversion, VkSamplerYcbcrRange, VkStructureType, vkCreateSamplerYcbcrConversion, vkCreateSamplerYcbcrConversionKHR

Constructors

SamplerYcbcrConversionCreateInfo

Fields

next :: Maybe SomeVkStruct

format :: Format

ycbcrModel :: SamplerYcbcrModelConversion

ycbcrRange :: SamplerYcbcrRange

components :: ComponentMapping

xChromaOffset :: ChromaLocation

yChromaOffset :: ChromaLocation

chromaFilter :: Filter

forceExplicitReconstruction :: Bool

withCStructSamplerYcbcrConversionImageFormatProperties :: SamplerYcbcrConversionImageFormatProperties -> (VkSamplerYcbcrConversionImageFormatProperties -> IO a) -> IO a

A function to temporarily allocate memory for a VkSamplerYcbcrConversionImageFormatProperties and marshal a SamplerYcbcrConversionImageFormatProperties into it. The VkSamplerYcbcrConversionImageFormatProperties is only valid inside the provided computation and must not be returned out of it.

fromCStructSamplerYcbcrConversionImageFormatProperties :: VkSamplerYcbcrConversionImageFormatProperties -> IO SamplerYcbcrConversionImageFormatProperties

A function to read a VkSamplerYcbcrConversionImageFormatProperties and all additional structures in the pointer chain into a SamplerYcbcrConversionImageFormatProperties.

data SamplerYcbcrConversionImageFormatProperties

VkSamplerYcbcrConversionImageFormatProperties - Structure specifying combined image sampler descriptor count for multi-planar images

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

SamplerYcbcrConversionImageFormatProperties

Fields

next :: Maybe SomeVkStruct

combinedImageSamplerDescriptorCount :: Word32

withCStructSamplerYcbcrConversionInfo :: SamplerYcbcrConversionInfo -> (VkSamplerYcbcrConversionInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkSamplerYcbcrConversionInfo and marshal a SamplerYcbcrConversionInfo into it. The VkSamplerYcbcrConversionInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructSamplerYcbcrConversionInfo :: VkSamplerYcbcrConversionInfo -> IO SamplerYcbcrConversionInfo

A function to read a VkSamplerYcbcrConversionInfo and all additional structures in the pointer chain into a SamplerYcbcrConversionInfo.

data SamplerYcbcrConversionInfo

VkSamplerYcbcrConversionInfo - Structure specifying Y’CbCr conversion to a sampler or image view

Valid Usage (Implicit)

See Also

VkSamplerYcbcrConversion, VkStructureType

Constructors

SamplerYcbcrConversionInfo

Fields

next :: Maybe SomeVkStruct

conversion :: SamplerYcbcrConversion

type SamplerYcbcrModelConversion = VkSamplerYcbcrModelConversion

VkSamplerYcbcrModelConversion - Color model component of a color space

Description

VK_SAMPLER_YCBCR_MODEL_CONVERSION_RGB_IDENTITY specifies that the input values to the conversion are unmodified.

VK_SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_IDENTITY specifies no model conversion but the inputs are range expanded as for Y’CBCR.

VK_SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_709 specifies the color model conversion from Y’CBCR to R’G’B' defined in BT.709 and described in the “BT.709 Y’CBCR conversion” section of the Khronos Data Format Specification.

VK_SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_601 specifies the color model conversion from Y’CBCR to R’G’B' defined in BT.601 and described in the “BT.601 Y’CBCR conversion” section of the Khronos Data Format Specification.

VK_SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_2020 specifies the color model conversion from Y’CBCR to R’G’B' defined in BT.2020 and described in the “BT.2020 Y’CBCR conversion” section of the Khronos Data Format Specification.

In the VK_SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_* color models, for the input to the sampler Y’CBCR range expansion and model conversion:

the Y (Y' luma) channel corresponds to the G channel of an RGB image.

the CB (CB or “U” blue color difference) channel corresponds to the B channel of an RGB image.

the CR (CR or “V” red color difference) channel corresponds to the R channel of an RGB image.

the alpha channel, if present, is not modified by color model conversion.

These rules reflect the mapping of channels after the channel swizzle operation (controlled by VkSamplerYcbcrConversionCreateInfo::components).

Note

For example, an “YUVA” 32-bit format comprising four 8-bit channels can be implemented as VK_FORMAT_R8G8B8A8_UNORM with a component mapping:

components.a = VK_COMPONENT_SWIZZLE_IDENTITY

components.r = VK_COMPONENT_SWIZZLE_B

components.g = VK_COMPONENT_SWIZZLE_R

components.b = VK_COMPONENT_SWIZZLE_G

See Also

VkAndroidHardwareBufferFormatPropertiesANDROID, VkSamplerYcbcrConversionCreateInfo

pattern SAMPLER_YCBCR_MODEL_CONVERSION_RGB_IDENTITY :: a ~ SamplerYcbcrModelConversion => a

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_IDENTITY :: a ~ SamplerYcbcrModelConversion => a

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_709 :: a ~ SamplerYcbcrModelConversion => a

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_601 :: a ~ SamplerYcbcrModelConversion => a

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_2020 :: a ~ SamplerYcbcrModelConversion => a

type SamplerYcbcrModelConversionKHR = SamplerYcbcrModelConversion

type SamplerYcbcrRange = VkSamplerYcbcrRange

VkSamplerYcbcrRange - Range of encoded values in a color space

Description

The formulae for these conversions is described in the Sampler Y’CBCR Range Expansion section of the Image Operations chapter.

No range modification takes place if ycbcrModel is VK_SAMPLER_YCBCR_MODEL_CONVERSION_RGB_IDENTITY; the ycbcrRange field of VkSamplerYcbcrConversionCreateInfo is ignored in this case.

See Also

VkAndroidHardwareBufferFormatPropertiesANDROID, VkSamplerYcbcrConversionCreateInfo

pattern SAMPLER_YCBCR_RANGE_ITU_FULL :: a ~ SamplerYcbcrRange => a

VK_SAMPLER_YCBCR_RANGE_ITU_FULL specifies that the full range of the encoded values are valid and interpreted according to the ITU “full range” quantization rules.

pattern SAMPLER_YCBCR_RANGE_ITU_NARROW :: a ~ SamplerYcbcrRange => a

VK_SAMPLER_YCBCR_RANGE_ITU_NARROW specifies that headroom and foot room are reserved in the numerical range of encoded values, and the remaining values are expanded according to the ITU “narrow range” quantization rules.

type SamplerYcbcrRangeKHR = SamplerYcbcrRange

createSamplerYcbcrConversion :: Device -> SamplerYcbcrConversionCreateInfo -> Maybe AllocationCallbacks -> IO SamplerYcbcrConversion

vkCreateSamplerYcbcrConversion - Create a new Ycbcr conversion

Parameters

device is the logical device that creates the sampler Y’CBCR conversion.

pCreateInfo is a pointer to an instance of the VkSamplerYcbcrConversionCreateInfo specifying the requested sampler Y’CBCR conversion.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pYcbcrConversion points to a VkSamplerYcbcrConversion handle in which the resulting sampler Y’CBCR conversion is returned.

Description

The interpretation of the configured sampler Y’CBCR conversion is described in more detail in the description of sampler Y’CBCR conversion in the Image Operations chapter.

Valid Usage

The sampler Y’CBCR conversion feature must be enabled

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkSamplerYcbcrConversionCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pYcbcrConversion must be a valid pointer to a VkSamplerYcbcrConversion handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDevice, VkSamplerYcbcrConversion, VkSamplerYcbcrConversionCreateInfo

destroySamplerYcbcrConversion :: Device -> SamplerYcbcrConversion -> Maybe AllocationCallbacks -> IO ()

vkDestroySamplerYcbcrConversion - Destroy a created Y’CbCr conversion

Parameters

device is the logical device that destroys the Y’CBCR conversion.

ycbcrConversion is the conversion to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage (Implicit)

device must be a valid VkDevice handle

If ycbcrConversion is not VK_NULL_HANDLE, ycbcrConversion must be a valid VkSamplerYcbcrConversion handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

If ycbcrConversion is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to ycbcrConversion must be externally synchronized

See Also

VkAllocationCallbacks, VkDevice, VkSamplerYcbcrConversion

withSamplerYcbcrConversion :: Device -> SamplerYcbcrConversionCreateInfo -> Maybe AllocationCallbacks -> (SamplerYcbcrConversion -> IO a) -> IO a

A safe wrapper for createSamplerYcbcrConversion and destroySamplerYcbcrConversion using bracket

The allocated value must not be returned from the provided computation

pattern STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_CREATE_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_BIND_IMAGE_PLANE_MEMORY_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_IMAGE_PLANE_MEMORY_REQUIREMENTS_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SAMPLER_YCBCR_CONVERSION_FEATURES :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_IMAGE_FORMAT_PROPERTIES :: a ~ StructureType => a

pattern OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION :: a ~ ObjectType => a

pattern FORMAT_G8B8G8R8_422_UNORM :: a ~ Format => a

pattern FORMAT_B8G8R8G8_422_UNORM :: a ~ Format => a

pattern FORMAT_G8_B8_R8_3PLANE_420_UNORM :: a ~ Format => a

pattern FORMAT_G8_B8R8_2PLANE_420_UNORM :: a ~ Format => a

pattern FORMAT_G8_B8_R8_3PLANE_422_UNORM :: a ~ Format => a

pattern FORMAT_G8_B8R8_2PLANE_422_UNORM :: a ~ Format => a

pattern FORMAT_G8_B8_R8_3PLANE_444_UNORM :: a ~ Format => a

pattern FORMAT_R10X6_UNORM_PACK16 :: a ~ Format => a

pattern FORMAT_R10X6G10X6_UNORM_2PACK16 :: a ~ Format => a

pattern FORMAT_R10X6G10X6B10X6A10X6_UNORM_4PACK16 :: a ~ Format => a

pattern FORMAT_G10X6B10X6G10X6R10X6_422_UNORM_4PACK16 :: a ~ Format => a

pattern FORMAT_B10X6G10X6R10X6G10X6_422_UNORM_4PACK16 :: a ~ Format => a

pattern FORMAT_G10X6_B10X6_R10X6_3PLANE_420_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G10X6_B10X6R10X6_2PLANE_420_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G10X6_B10X6_R10X6_3PLANE_422_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G10X6_B10X6R10X6_2PLANE_422_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G10X6_B10X6_R10X6_3PLANE_444_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_R12X4_UNORM_PACK16 :: a ~ Format => a

pattern FORMAT_R12X4G12X4_UNORM_2PACK16 :: a ~ Format => a

pattern FORMAT_R12X4G12X4B12X4A12X4_UNORM_4PACK16 :: a ~ Format => a

pattern FORMAT_G12X4B12X4G12X4R12X4_422_UNORM_4PACK16 :: a ~ Format => a

pattern FORMAT_B12X4G12X4R12X4G12X4_422_UNORM_4PACK16 :: a ~ Format => a

pattern FORMAT_G12X4_B12X4_R12X4_3PLANE_420_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G12X4_B12X4R12X4_2PLANE_420_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G12X4_B12X4_R12X4_3PLANE_422_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G12X4_B12X4R12X4_2PLANE_422_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G12X4_B12X4_R12X4_3PLANE_444_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G16B16G16R16_422_UNORM :: a ~ Format => a

pattern FORMAT_B16G16R16G16_422_UNORM :: a ~ Format => a

pattern FORMAT_G16_B16_R16_3PLANE_420_UNORM :: a ~ Format => a

pattern FORMAT_G16_B16R16_2PLANE_420_UNORM :: a ~ Format => a

pattern FORMAT_G16_B16_R16_3PLANE_422_UNORM :: a ~ Format => a

pattern FORMAT_G16_B16R16_2PLANE_422_UNORM :: a ~ Format => a

pattern FORMAT_G16_B16_R16_3PLANE_444_UNORM :: a ~ Format => a

pattern IMAGE_ASPECT_PLANE_0_BIT :: a ~ ImageAspectFlagBits => a

pattern IMAGE_ASPECT_PLANE_1_BIT :: a ~ ImageAspectFlagBits => a

pattern IMAGE_ASPECT_PLANE_2_BIT :: a ~ ImageAspectFlagBits => a

pattern IMAGE_CREATE_DISJOINT_BIT :: a ~ ImageCreateFlagBits => a

pattern FORMAT_FEATURE_MIDPOINT_CHROMA_SAMPLES_BIT :: a ~ FormatFeatureFlagBits => a

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_LINEAR_FILTER_BIT :: a ~ FormatFeatureFlagBits => a

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_SEPARATE_RECONSTRUCTION_FILTER_BIT :: a ~ FormatFeatureFlagBits => a

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_BIT :: a ~ FormatFeatureFlagBits => a

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_FORCEABLE_BIT :: a ~ FormatFeatureFlagBits => a

pattern FORMAT_FEATURE_DISJOINT_BIT :: a ~ FormatFeatureFlagBits => a

pattern FORMAT_FEATURE_COSITED_CHROMA_SAMPLES_BIT :: a ~ FormatFeatureFlagBits => a