Graphics.Vulkan.Core11.Promoted_from_VK_KHR_shader_draw_parameters

Documentation

type PhysicalDeviceShaderDrawParameterFeatures = PhysicalDeviceShaderDrawParametersFeatures

withCStructPhysicalDeviceShaderDrawParametersFeatures :: PhysicalDeviceShaderDrawParametersFeatures -> (VkPhysicalDeviceShaderDrawParametersFeatures -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceShaderDrawParametersFeatures and marshal a PhysicalDeviceShaderDrawParametersFeatures into it. The VkPhysicalDeviceShaderDrawParametersFeatures is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceShaderDrawParametersFeatures :: VkPhysicalDeviceShaderDrawParametersFeatures -> IO PhysicalDeviceShaderDrawParametersFeatures

A function to read a VkPhysicalDeviceShaderDrawParametersFeatures and all additional structures in the pointer chain into a PhysicalDeviceShaderDrawParametersFeatures.

data PhysicalDeviceShaderDrawParametersFeatures

VkPhysicalDeviceShaderDrawParametersFeatures - Structure describing shader draw parameter features that can be supported by an implementation

Description

If the VkPhysicalDeviceShaderDrawParametersFeatures structure is included in the pNext chain of VkPhysicalDeviceFeatures2, it is filled with a value indicating whether the feature is supported.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceShaderDrawParametersFeatures

Fields

next :: Maybe SomeVkStruct

shaderDrawParameters :: Bool

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_DRAW_PARAMETER_FEATURES :: VkStructureType

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_DRAW_PARAMETERS_FEATURES :: a ~ StructureType => a