Graphics.Vulkan.Core12.Enums.ResolveModeFlagBits

Documentation

newtype ResolveModeFlagBits

VkResolveModeFlagBits - Bitmask indicating supported depth and stencil resolve modes

See Also

ResolveModeFlags, SubpassDescriptionDepthStencilResolve

Constructors

ResolveModeFlagBits Flags

Bundled Patterns

pattern RESOLVE_MODE_NONE :: ResolveModeFlagBits
RESOLVE_MODE_NONE indicates that no resolve operation is done.

pattern RESOLVE_MODE_SAMPLE_ZERO_BIT :: ResolveModeFlagBits
RESOLVE_MODE_SAMPLE_ZERO_BIT indicates that result of the resolve operation is equal to the value of sample 0.

pattern RESOLVE_MODE_AVERAGE_BIT :: ResolveModeFlagBits
RESOLVE_MODE_AVERAGE_BIT indicates that result of the resolve operation is the average of the sample values.

pattern RESOLVE_MODE_MIN_BIT :: ResolveModeFlagBits
RESOLVE_MODE_MIN_BIT indicates that result of the resolve operation is the minimum of the sample values.

pattern RESOLVE_MODE_MAX_BIT :: ResolveModeFlagBits
RESOLVE_MODE_MAX_BIT indicates that result of the resolve operation is the maximum of the sample values.

type ResolveModeFlags = ResolveModeFlagBits