Graphics.Vulkan.Core12.Enums.SemaphoreType

Documentation

newtype SemaphoreType

VkSemaphoreType - Sepcifies the type of a semaphore object

See Also

SemaphoreTypeCreateInfo

Constructors

SemaphoreType Int32

Bundled Patterns

pattern SEMAPHORE_TYPE_BINARY :: SemaphoreType
SEMAPHORE_TYPE_BINARY specifies a binary semaphore type that has a boolean payload indicating whether the semaphore is currently signaled or unsignaled. When created, the semaphore is in the unsignaled state.

pattern SEMAPHORE_TYPE_TIMELINE :: SemaphoreType
SEMAPHORE_TYPE_TIMELINE specifies a timeline semaphore type that has a monotonically increasing 64-bit unsigned integer payload indicating whether the semaphore is signaled with respect to a particular reference value. When created, the semaphore payload has the value given by the initialValue field of SemaphoreTypeCreateInfo.