Graphics.Vulkan.Extensions.Handles

Documentation

newtype ObjectTableNVX

VkObjectTableNVX - Opaque handle to an object table

See Also

CmdProcessCommandsInfoNVX, CmdReserveSpaceForCommandsInfoNVX, createObjectTableNVX, destroyObjectTableNVX, registerObjectsNVX, unregisterObjectsNVX

Constructors

ObjectTableNVX Word64

newtype IndirectCommandsLayoutNVX

VkIndirectCommandsLayoutNVX - Opaque handle to an indirect commands layout object

See Also

CmdProcessCommandsInfoNVX, CmdReserveSpaceForCommandsInfoNVX, createIndirectCommandsLayoutNVX, destroyIndirectCommandsLayoutNVX

Constructors

IndirectCommandsLayoutNVX Word64

newtype ValidationCacheEXT

VkValidationCacheEXT - Opaque handle to a validation cache object

See Also

ShaderModuleValidationCacheCreateInfoEXT, createValidationCacheEXT, destroyValidationCacheEXT, getValidationCacheDataEXT, mergeValidationCachesEXT

Constructors

ValidationCacheEXT Word64

newtype AccelerationStructureNV

VkAccelerationStructureNV - Opaque handle to an acceleration structure object

See Also

AccelerationStructureMemoryRequirementsInfoNV, BindAccelerationStructureMemoryInfoNV, WriteDescriptorSetAccelerationStructureNV, cmdBuildAccelerationStructureNV, cmdCopyAccelerationStructureNV, cmdWriteAccelerationStructuresPropertiesNV, createAccelerationStructureNV, destroyAccelerationStructureNV, getAccelerationStructureHandleNV

Constructors

AccelerationStructureNV Word64

newtype PerformanceConfigurationINTEL

VkPerformanceConfigurationINTEL - Device configuration for performance queries

See Also

acquirePerformanceConfigurationINTEL, queueSetPerformanceConfigurationINTEL, releasePerformanceConfigurationINTEL

Constructors

PerformanceConfigurationINTEL Word64

newtype DisplayKHR

VkDisplayKHR - Opaque handle to a display object

See Also

DisplayPlanePropertiesKHR, DisplayPropertiesKHR, acquireXlibDisplayEXT, createDisplayModeKHR, displayPowerControlEXT, getDisplayModeProperties2KHR, getDisplayModePropertiesKHR, getDisplayPlaneSupportedDisplaysKHR, getRandROutputDisplayEXT, registerDisplayEventEXT, releaseDisplayEXT

Constructors

DisplayKHR Word64

newtype DisplayModeKHR

VkDisplayModeKHR - Opaque handle to a display mode object

See Also

DisplayModePropertiesKHR, DisplayPlaneInfo2KHR, DisplaySurfaceCreateInfoKHR, createDisplayModeKHR, getDisplayPlaneCapabilitiesKHR

Constructors

DisplayModeKHR Word64

newtype SurfaceKHR

VkSurfaceKHR - Opaque handle to a surface object

Description

The VK_KHR_surface extension declares the SurfaceKHR object, and provides a function for destroying SurfaceKHR objects. Separate platform-specific extensions each provide a function for creating a SurfaceKHR object for the respective platform. From the application’s perspective this is an opaque handle, just like the handles of other Vulkan objects.

See Also

PhysicalDeviceSurfaceInfo2KHR, SwapchainCreateInfoKHR, createAndroidSurfaceKHR, createDisplayPlaneSurfaceKHR, createHeadlessSurfaceEXT, createIOSSurfaceMVK, createImagePipeSurfaceFUCHSIA, createMacOSSurfaceMVK, createMetalSurfaceEXT, createStreamDescriptorSurfaceGGP, createViSurfaceNN, createWaylandSurfaceKHR, createWin32SurfaceKHR, createXcbSurfaceKHR, createXlibSurfaceKHR, destroySurfaceKHR, getDeviceGroupSurfacePresentModesKHR, getPhysicalDevicePresentRectanglesKHR, getPhysicalDeviceSurfaceCapabilities2EXT, getPhysicalDeviceSurfaceCapabilitiesKHR, getPhysicalDeviceSurfaceFormatsKHR, getPhysicalDeviceSurfacePresentModesKHR, getPhysicalDeviceSurfaceSupportKHR

Constructors

SurfaceKHR Word64

newtype SwapchainKHR

VkSwapchainKHR - Opaque handle to a swapchain object

Description

A swapchain is an abstraction for an array of presentable images that are associated with a surface. The presentable images are represented by Image objects created by the platform. One image (which can be an array image for multiview/stereoscopic-3D surfaces) is displayed at a time, but multiple images can be queued for presentation. An application renders to the image, and then queues the image for presentation to the surface.

A native window cannot be associated with more than one non-retired swapchain at a time. Further, swapchains cannot be created for native windows that have a non-Vulkan graphics API surface associated with them.

Note

The presentation engine is an abstraction for the platform’s compositor or display engine.

The presentation engine may be synchronous or asynchronous with respect to the application and/or logical device.

Some implementations may use the device’s graphics queue or dedicated presentation hardware to perform presentation.

The presentable images of a swapchain are owned by the presentation engine. An application can acquire use of a presentable image from the presentation engine. Use of a presentable image must occur only after the image is returned by acquireNextImageKHR, and before it is presented by queuePresentKHR. This includes transitioning the image layout and rendering commands.

An application can acquire use of a presentable image with acquireNextImageKHR. After acquiring a presentable image and before modifying it, the application must use a synchronization primitive to ensure that the presentation engine has finished reading from the image. The application can then transition the image’s layout, queue rendering commands to it, etc. Finally, the application presents the image with queuePresentKHR, which releases the acquisition of the image.

The presentation engine controls the order in which presentable images are acquired for use by the application.

Note

This allows the platform to handle situations which require out-of-order return of images after presentation. At the same time, it allows the application to generate command buffers referencing all of the images in the swapchain at initialization time, rather than in its main loop.

See Also

AcquireNextImageInfoKHR, BindImageMemorySwapchainInfoKHR, ImageSwapchainCreateInfoKHR, PresentInfoKHR, SwapchainCreateInfoKHR, acquireFullScreenExclusiveModeEXT, acquireNextImageKHR, createSharedSwapchainsKHR, createSwapchainKHR, destroySwapchainKHR, getPastPresentationTimingGOOGLE, getRefreshCycleDurationGOOGLE, getSwapchainCounterEXT, getSwapchainImagesKHR, getSwapchainStatusKHR, releaseFullScreenExclusiveModeEXT, setHdrMetadataEXT, setLocalDimmingAMD

Constructors

SwapchainKHR Word64

newtype DebugReportCallbackEXT

VkDebugReportCallbackEXT - Opaque handle to a debug report callback object

See Also

createDebugReportCallbackEXT, destroyDebugReportCallbackEXT

Constructors

DebugReportCallbackEXT Word64

newtype DebugUtilsMessengerEXT

VkDebugUtilsMessengerEXT - Opaque handle to a debug messenger object

Description

The debug messenger will provide detailed feedback on the application’s use of Vulkan when events of interest occur. When an event of interest does occur, the debug messenger will submit a debug message to the debug callback that was provided during its creation. Additionally, the debug messenger is responsible with filtering out debug messages that the callback is not interested in and will only provide desired debug messages.

See Also

createDebugUtilsMessengerEXT, destroyDebugUtilsMessengerEXT

Constructors

DebugUtilsMessengerEXT Word64

data Instance

VkInstance - Opaque handle to an instance object

See Also

createAndroidSurfaceKHR, createDebugReportCallbackEXT, createDebugUtilsMessengerEXT, createDisplayPlaneSurfaceKHR, createHeadlessSurfaceEXT, createIOSSurfaceMVK, createImagePipeSurfaceFUCHSIA, createInstance, createMacOSSurfaceMVK, createMetalSurfaceEXT, createStreamDescriptorSurfaceGGP, createViSurfaceNN, createWaylandSurfaceKHR, createWin32SurfaceKHR, createXcbSurfaceKHR, createXlibSurfaceKHR, debugReportMessageEXT, destroyDebugReportCallbackEXT, destroyDebugUtilsMessengerEXT, destroyInstance, destroySurfaceKHR, enumeratePhysicalDeviceGroups, enumeratePhysicalDeviceGroupsKHR, enumeratePhysicalDevices, getInstanceProcAddr, submitDebugUtilsMessageEXT

Constructors

Instance

Fields

instanceHandle :: Ptr Instance_T

instanceCmds :: InstanceCmds

data PhysicalDevice

VkPhysicalDevice - Opaque handle to a physical device object

See Also

DeviceGroupDeviceCreateInfo, PhysicalDeviceGroupProperties, acquireXlibDisplayEXT, createDevice, createDisplayModeKHR, enumerateDeviceExtensionProperties, enumerateDeviceLayerProperties, enumeratePhysicalDeviceQueueFamilyPerformanceQueryCountersKHR, enumeratePhysicalDevices, getDisplayModeProperties2KHR, getDisplayModePropertiesKHR, getDisplayPlaneCapabilities2KHR, getDisplayPlaneCapabilitiesKHR, getDisplayPlaneSupportedDisplaysKHR, getPhysicalDeviceCalibrateableTimeDomainsEXT, getPhysicalDeviceCooperativeMatrixPropertiesNV, getPhysicalDeviceDisplayPlaneProperties2KHR, getPhysicalDeviceDisplayPlanePropertiesKHR, getPhysicalDeviceDisplayProperties2KHR, getPhysicalDeviceDisplayPropertiesKHR, getPhysicalDeviceExternalBufferProperties, getPhysicalDeviceExternalBufferPropertiesKHR, getPhysicalDeviceExternalFenceProperties, getPhysicalDeviceExternalFencePropertiesKHR, getPhysicalDeviceExternalImageFormatPropertiesNV, getPhysicalDeviceExternalSemaphoreProperties, getPhysicalDeviceExternalSemaphorePropertiesKHR, getPhysicalDeviceFeatures, getPhysicalDeviceFeatures2, getPhysicalDeviceFeatures2KHR, getPhysicalDeviceFormatProperties, getPhysicalDeviceFormatProperties2, getPhysicalDeviceFormatProperties2KHR, getPhysicalDeviceGeneratedCommandsPropertiesNVX, getPhysicalDeviceImageFormatProperties, getPhysicalDeviceImageFormatProperties2, getPhysicalDeviceImageFormatProperties2KHR, getPhysicalDeviceMemoryProperties, getPhysicalDeviceMemoryProperties2, getPhysicalDeviceMemoryProperties2KHR, getPhysicalDeviceMultisamplePropertiesEXT, getPhysicalDevicePresentRectanglesKHR, getPhysicalDeviceProperties, getPhysicalDeviceProperties2, getPhysicalDeviceProperties2KHR, getPhysicalDeviceQueueFamilyPerformanceQueryPassesKHR, getPhysicalDeviceQueueFamilyProperties, getPhysicalDeviceQueueFamilyProperties2, getPhysicalDeviceQueueFamilyProperties2KHR, getPhysicalDeviceSparseImageFormatProperties, getPhysicalDeviceSparseImageFormatProperties2, getPhysicalDeviceSparseImageFormatProperties2KHR, getPhysicalDeviceSupportedFramebufferMixedSamplesCombinationsNV, getPhysicalDeviceSurfaceCapabilities2EXT, getPhysicalDeviceSurfaceCapabilities2KHR, getPhysicalDeviceSurfaceCapabilitiesKHR, getPhysicalDeviceSurfaceFormats2KHR, getPhysicalDeviceSurfaceFormatsKHR, getPhysicalDeviceSurfacePresentModes2EXT, getPhysicalDeviceSurfacePresentModesKHR, getPhysicalDeviceSurfaceSupportKHR, getPhysicalDeviceToolPropertiesEXT, getPhysicalDeviceWaylandPresentationSupportKHR, getPhysicalDeviceWin32PresentationSupportKHR, getPhysicalDeviceXcbPresentationSupportKHR, getPhysicalDeviceXlibPresentationSupportKHR, getRandROutputDisplayEXT, releaseDisplayEXT

Constructors

PhysicalDevice

Fields

physicalDeviceHandle :: Ptr PhysicalDevice_T

instanceCmds :: InstanceCmds

data Device

VkDevice - Opaque handle to a device object

See Also

acquireFullScreenExclusiveModeEXT, acquireNextImage2KHR, acquireNextImageKHR, acquirePerformanceConfigurationINTEL, acquireProfilingLockKHR, allocateCommandBuffers, allocateDescriptorSets, allocateMemory, bindAccelerationStructureMemoryNV, bindBufferMemory, bindBufferMemory2, bindBufferMemory2KHR, bindImageMemory, bindImageMemory2, bindImageMemory2KHR, compileDeferredNV, createAccelerationStructureNV, createBuffer, createBufferView, createCommandPool, createComputePipelines, createDescriptorPool, createDescriptorSetLayout, createDescriptorUpdateTemplate, createDescriptorUpdateTemplateKHR, createDevice, createEvent, createFence, createFramebuffer, createGraphicsPipelines, createImage, createImageView, createIndirectCommandsLayoutNVX, createObjectTableNVX, createPipelineCache, createPipelineLayout, createQueryPool, createRayTracingPipelinesNV, createRenderPass, createRenderPass2, createRenderPass2KHR, createSampler, createSamplerYcbcrConversion, createSamplerYcbcrConversionKHR, createSemaphore, createShaderModule, createSharedSwapchainsKHR, createSwapchainKHR, createValidationCacheEXT, debugMarkerSetObjectNameEXT, debugMarkerSetObjectTagEXT, destroyAccelerationStructureNV, destroyBuffer, destroyBufferView, destroyCommandPool, destroyDescriptorPool, destroyDescriptorSetLayout, destroyDescriptorUpdateTemplate, destroyDescriptorUpdateTemplateKHR, destroyDevice, destroyEvent, destroyFence, destroyFramebuffer, destroyImage, destroyImageView, destroyIndirectCommandsLayoutNVX, destroyObjectTableNVX, destroyPipeline, destroyPipelineCache, destroyPipelineLayout, destroyQueryPool, destroyRenderPass, destroySampler, destroySamplerYcbcrConversion, destroySamplerYcbcrConversionKHR, destroySemaphore, destroyShaderModule, destroySwapchainKHR, destroyValidationCacheEXT, deviceWaitIdle, displayPowerControlEXT, flushMappedMemoryRanges, freeCommandBuffers, freeDescriptorSets, freeMemory, getAccelerationStructureHandleNV, getAccelerationStructureMemoryRequirementsNV, getAndroidHardwareBufferPropertiesANDROID, getBufferDeviceAddress, getBufferDeviceAddressEXT, getBufferDeviceAddressKHR, getBufferMemoryRequirements, getBufferMemoryRequirements2, getBufferMemoryRequirements2KHR, getBufferOpaqueCaptureAddress, getBufferOpaqueCaptureAddressKHR, getCalibratedTimestampsEXT, getDescriptorSetLayoutSupport, getDescriptorSetLayoutSupportKHR, getDeviceGroupPeerMemoryFeatures, getDeviceGroupPeerMemoryFeaturesKHR, getDeviceGroupPresentCapabilitiesKHR, getDeviceGroupSurfacePresentModes2EXT, getDeviceGroupSurfacePresentModesKHR, getDeviceMemoryCommitment, getDeviceMemoryOpaqueCaptureAddress, getDeviceMemoryOpaqueCaptureAddressKHR, getDeviceProcAddr, getDeviceQueue, getDeviceQueue2, getEventStatus, getFenceFdKHR, getFenceStatus, getFenceWin32HandleKHR, getImageDrmFormatModifierPropertiesEXT, getImageMemoryRequirements, getImageMemoryRequirements2, getImageMemoryRequirements2KHR, getImageSparseMemoryRequirements, getImageSparseMemoryRequirements2, getImageSparseMemoryRequirements2KHR, getImageSubresourceLayout, getImageViewHandleNVX, getMemoryAndroidHardwareBufferANDROID, getMemoryFdKHR, getMemoryFdPropertiesKHR, getMemoryHostPointerPropertiesEXT, getMemoryWin32HandleKHR, getMemoryWin32HandleNV, getMemoryWin32HandlePropertiesKHR, getPastPresentationTimingGOOGLE, getPerformanceParameterINTEL, getPipelineCacheData, getPipelineExecutableInternalRepresentationsKHR, getPipelineExecutablePropertiesKHR, getPipelineExecutableStatisticsKHR, getQueryPoolResults, getRayTracingShaderGroupHandlesNV, getRefreshCycleDurationGOOGLE, getRenderAreaGranularity, getSemaphoreCounterValue, getSemaphoreCounterValueKHR, getSemaphoreFdKHR, getSemaphoreWin32HandleKHR, getShaderInfoAMD, getSwapchainCounterEXT, getSwapchainImagesKHR, getSwapchainStatusKHR, getValidationCacheDataEXT, importFenceFdKHR, importFenceWin32HandleKHR, importSemaphoreFdKHR, importSemaphoreWin32HandleKHR, initializePerformanceApiINTEL, invalidateMappedMemoryRanges, mapMemory, mergePipelineCaches, mergeValidationCachesEXT, registerDeviceEventEXT, registerDisplayEventEXT, registerObjectsNVX, releaseFullScreenExclusiveModeEXT, releasePerformanceConfigurationINTEL, releaseProfilingLockKHR, resetCommandPool, resetDescriptorPool, resetEvent, resetFences, resetQueryPool, resetQueryPoolEXT, setDebugUtilsObjectNameEXT, setDebugUtilsObjectTagEXT, setEvent, setHdrMetadataEXT, setLocalDimmingAMD, signalSemaphore, signalSemaphoreKHR, trimCommandPool, trimCommandPoolKHR, uninitializePerformanceApiINTEL, unmapMemory, unregisterObjectsNVX, updateDescriptorSetWithTemplate, updateDescriptorSetWithTemplateKHR, updateDescriptorSets, waitForFences, waitSemaphores, waitSemaphoresKHR

Constructors

Device

Fields

deviceHandle :: Ptr Device_T

deviceCmds :: DeviceCmds

data Queue

VkQueue - Opaque handle to a queue object

See Also

getDeviceQueue, getDeviceQueue2, getQueueCheckpointDataNV, queueBeginDebugUtilsLabelEXT, queueBindSparse, queueEndDebugUtilsLabelEXT, queueInsertDebugUtilsLabelEXT, queuePresentKHR, queueSetPerformanceConfigurationINTEL, queueSubmit, queueWaitIdle

Constructors

Queue

Fields

queueHandle :: Ptr Queue_T

deviceCmds :: DeviceCmds

data CommandBuffer

VkCommandBuffer - Opaque handle to a command buffer object

See Also

CmdProcessCommandsInfoNVX, SubmitInfo, allocateCommandBuffers, beginCommandBuffer, cmdBeginConditionalRenderingEXT, cmdBeginDebugUtilsLabelEXT, cmdBeginQuery, cmdBeginQueryIndexedEXT, cmdBeginRenderPass, cmdBeginRenderPass2, cmdBeginRenderPass2KHR, cmdBeginTransformFeedbackEXT, cmdBindDescriptorSets, cmdBindIndexBuffer, cmdBindPipeline, cmdBindShadingRateImageNV, cmdBindTransformFeedbackBuffersEXT, cmdBindVertexBuffers, cmdBlitImage, cmdBuildAccelerationStructureNV, cmdClearAttachments, cmdClearColorImage, cmdClearDepthStencilImage, cmdCopyAccelerationStructureNV, cmdCopyBuffer, cmdCopyBufferToImage, cmdCopyImage, cmdCopyImageToBuffer, cmdCopyQueryPoolResults, cmdDebugMarkerBeginEXT, cmdDebugMarkerEndEXT, cmdDebugMarkerInsertEXT, cmdDispatch, cmdDispatchBase, cmdDispatchBaseKHR, cmdDispatchIndirect, cmdDraw, cmdDrawIndexed, cmdDrawIndexedIndirect, cmdDrawIndexedIndirectCount, cmdDrawIndexedIndirectCountAMD, cmdDrawIndexedIndirectCountKHR, cmdDrawIndirect, cmdDrawIndirectByteCountEXT, cmdDrawIndirectCount, cmdDrawIndirectCountAMD, cmdDrawIndirectCountKHR, cmdDrawMeshTasksIndirectCountNV, cmdDrawMeshTasksIndirectNV, cmdDrawMeshTasksNV, cmdEndConditionalRenderingEXT, cmdEndDebugUtilsLabelEXT, cmdEndQuery, cmdEndQueryIndexedEXT, cmdEndRenderPass, cmdEndRenderPass2, cmdEndRenderPass2KHR, cmdEndTransformFeedbackEXT, cmdExecuteCommands, cmdFillBuffer, cmdInsertDebugUtilsLabelEXT, cmdNextSubpass, cmdNextSubpass2, cmdNextSubpass2KHR, cmdPipelineBarrier, cmdProcessCommandsNVX, cmdPushConstants, cmdPushDescriptorSetKHR, cmdPushDescriptorSetWithTemplateKHR, cmdReserveSpaceForCommandsNVX, cmdResetEvent, cmdResetQueryPool, cmdResolveImage, cmdSetBlendConstants, cmdSetCheckpointNV, cmdSetCoarseSampleOrderNV, cmdSetDepthBias, cmdSetDepthBounds, cmdSetDeviceMask, cmdSetDeviceMaskKHR, cmdSetDiscardRectangleEXT, cmdSetEvent, cmdSetExclusiveScissorNV, cmdSetLineStippleEXT, cmdSetLineWidth, cmdSetPerformanceMarkerINTEL, cmdSetPerformanceOverrideINTEL, cmdSetPerformanceStreamMarkerINTEL, cmdSetSampleLocationsEXT, cmdSetScissor, cmdSetStencilCompareMask, cmdSetStencilReference, cmdSetStencilWriteMask, cmdSetViewport, cmdSetViewportShadingRatePaletteNV, cmdSetViewportWScalingNV, cmdTraceRaysNV, cmdUpdateBuffer, cmdWaitEvents, cmdWriteAccelerationStructuresPropertiesNV, cmdWriteBufferMarkerAMD, cmdWriteTimestamp, endCommandBuffer, freeCommandBuffers, resetCommandBuffer

Constructors

CommandBuffer

Fields

commandBufferHandle :: Ptr CommandBuffer_T

deviceCmds :: DeviceCmds

newtype DeviceMemory

VkDeviceMemory - Opaque handle to a device memory object

See Also

BindAccelerationStructureMemoryInfoNV, BindBufferMemoryInfo, BindImageMemoryInfo, DeviceMemoryOpaqueCaptureAddressInfo, MappedMemoryRange, MemoryGetAndroidHardwareBufferInfoANDROID, MemoryGetFdInfoKHR, MemoryGetWin32HandleInfoKHR, SparseImageMemoryBind, SparseMemoryBind, Win32KeyedMutexAcquireReleaseInfoKHR, Win32KeyedMutexAcquireReleaseInfoNV, allocateMemory, bindBufferMemory, bindImageMemory, freeMemory, getDeviceMemoryCommitment, getMemoryWin32HandleNV, mapMemory, unmapMemory

Constructors

DeviceMemory Word64

newtype CommandPool

VkCommandPool - Opaque handle to a command pool object

See Also

CommandBufferAllocateInfo, createCommandPool, destroyCommandPool, freeCommandBuffers, resetCommandPool, trimCommandPool, trimCommandPoolKHR

Constructors

CommandPool Word64

newtype Buffer

VkBuffer - Opaque handle to a buffer object

See Also

BindBufferMemoryInfo, BufferDeviceAddressInfo, BufferMemoryBarrier, BufferMemoryRequirementsInfo2, BufferViewCreateInfo, CmdProcessCommandsInfoNVX, ConditionalRenderingBeginInfoEXT, DedicatedAllocationMemoryAllocateInfoNV, DescriptorBufferInfo, GeometryAABBNV, GeometryTrianglesNV, IndirectCommandsTokenNVX, MemoryDedicatedAllocateInfo, ObjectTableIndexBufferEntryNVX, ObjectTableVertexBufferEntryNVX, SparseBufferMemoryBindInfo, bindBufferMemory, cmdBeginTransformFeedbackEXT, cmdBindIndexBuffer, cmdBindTransformFeedbackBuffersEXT, cmdBindVertexBuffers, cmdBuildAccelerationStructureNV, cmdCopyBuffer, cmdCopyBufferToImage, cmdCopyImageToBuffer, cmdCopyQueryPoolResults, cmdDispatchIndirect, cmdDrawIndexedIndirect, cmdDrawIndexedIndirectCount, cmdDrawIndexedIndirectCountAMD, cmdDrawIndexedIndirectCountKHR, cmdDrawIndirect, cmdDrawIndirectByteCountEXT, cmdDrawIndirectCount, cmdDrawIndirectCountAMD, cmdDrawIndirectCountKHR, cmdDrawMeshTasksIndirectCountNV, cmdDrawMeshTasksIndirectNV, cmdEndTransformFeedbackEXT, cmdFillBuffer, cmdTraceRaysNV, cmdUpdateBuffer, cmdWriteBufferMarkerAMD, createBuffer, destroyBuffer, getBufferMemoryRequirements

Constructors

Buffer Word64

newtype BufferView

VkBufferView - Opaque handle to a buffer view object

See Also

WriteDescriptorSet, createBufferView, destroyBufferView

Constructors

BufferView Word64

newtype Image

VkImage - Opaque handle to an image object

See Also

BindImageMemoryInfo, DedicatedAllocationMemoryAllocateInfoNV, ImageMemoryBarrier, ImageMemoryRequirementsInfo2, ImageSparseMemoryRequirementsInfo2, ImageViewCreateInfo, MemoryDedicatedAllocateInfo, SparseImageMemoryBindInfo, SparseImageOpaqueMemoryBindInfo, bindImageMemory, cmdBlitImage, cmdClearColorImage, cmdClearDepthStencilImage, cmdCopyBufferToImage, cmdCopyImage, cmdCopyImageToBuffer, cmdResolveImage, createImage, destroyImage, getImageDrmFormatModifierPropertiesEXT, getImageMemoryRequirements, getImageSparseMemoryRequirements, getImageSubresourceLayout, getSwapchainImagesKHR

Constructors

Image Word64

newtype ImageView

VkImageView - Opaque handle to an image view object

See Also

DescriptorImageInfo, FramebufferCreateInfo, ImageViewHandleInfoNVX, RenderPassAttachmentBeginInfo, cmdBindShadingRateImageNV, createImageView, destroyImageView

Constructors

ImageView Word64

newtype ShaderModule

VkShaderModule - Opaque handle to a shader module object

See Also

PipelineShaderStageCreateInfo, createShaderModule, destroyShaderModule

Constructors

ShaderModule Word64

newtype Pipeline

VkPipeline - Opaque handle to a pipeline object

See Also

ComputePipelineCreateInfo, GraphicsPipelineCreateInfo, ObjectTablePipelineEntryNVX, PipelineExecutableInfoKHR, PipelineInfoKHR, RayTracingPipelineCreateInfoNV, cmdBindPipeline, compileDeferredNV, createComputePipelines, createGraphicsPipelines, createRayTracingPipelinesNV, destroyPipeline, getRayTracingShaderGroupHandlesNV, getShaderInfoAMD

Constructors

Pipeline Word64

newtype PipelineLayout

VkPipelineLayout - Opaque handle to a pipeline layout object

See Also

ComputePipelineCreateInfo, DescriptorUpdateTemplateCreateInfo, GraphicsPipelineCreateInfo, ObjectTableDescriptorSetEntryNVX, ObjectTablePushConstantEntryNVX, RayTracingPipelineCreateInfoNV, cmdBindDescriptorSets, cmdPushConstants, cmdPushDescriptorSetKHR, cmdPushDescriptorSetWithTemplateKHR, createPipelineLayout, destroyPipelineLayout

Constructors

PipelineLayout Word64

newtype Sampler

VkSampler - Opaque handle to a sampler object

See Also

DescriptorImageInfo, DescriptorSetLayoutBinding, ImageViewHandleInfoNVX, createSampler, destroySampler

Constructors

Sampler Word64

newtype DescriptorSet

VkDescriptorSet - Opaque handle to a descriptor set object

See Also

CopyDescriptorSet, ObjectTableDescriptorSetEntryNVX, WriteDescriptorSet, allocateDescriptorSets, cmdBindDescriptorSets, freeDescriptorSets, updateDescriptorSetWithTemplate, updateDescriptorSetWithTemplateKHR

Constructors

DescriptorSet Word64

newtype DescriptorSetLayout

VkDescriptorSetLayout - Opaque handle to a descriptor set layout object

See Also

DescriptorSetAllocateInfo, DescriptorUpdateTemplateCreateInfo, PipelineLayoutCreateInfo, createDescriptorSetLayout, destroyDescriptorSetLayout

Constructors

DescriptorSetLayout Word64

newtype Fence

VkFence - Opaque handle to a fence object

See Also

AcquireNextImageInfoKHR, FenceGetFdInfoKHR, FenceGetWin32HandleInfoKHR, ImportFenceFdInfoKHR, ImportFenceWin32HandleInfoKHR, acquireNextImageKHR, createFence, destroyFence, getFenceStatus, queueBindSparse, queueSubmit, registerDeviceEventEXT, registerDisplayEventEXT, resetFences, waitForFences

Constructors

Fence Word64

newtype Semaphore

VkSemaphore - Opaque handle to a semaphore object

See Also

AcquireNextImageInfoKHR, BindSparseInfo, ImportSemaphoreFdInfoKHR, ImportSemaphoreWin32HandleInfoKHR, PresentInfoKHR, SemaphoreGetFdInfoKHR, SemaphoreGetWin32HandleInfoKHR, SemaphoreSignalInfo, SemaphoreWaitInfo, SubmitInfo, acquireNextImageKHR, createSemaphore, destroySemaphore, getSemaphoreCounterValue, getSemaphoreCounterValueKHR

Constructors

Semaphore Word64

newtype QueryPool

VkQueryPool - Opaque handle to a query pool object

See Also

cmdBeginQuery, cmdBeginQueryIndexedEXT, cmdCopyQueryPoolResults, cmdEndQuery, cmdEndQueryIndexedEXT, cmdResetQueryPool, cmdWriteAccelerationStructuresPropertiesNV, cmdWriteTimestamp, createQueryPool, destroyQueryPool, getQueryPoolResults, resetQueryPool, resetQueryPoolEXT

Constructors

QueryPool Word64

newtype Framebuffer

VkFramebuffer - Opaque handle to a framebuffer object

See Also

CommandBufferInheritanceInfo, RenderPassBeginInfo, createFramebuffer, destroyFramebuffer

Constructors

Framebuffer Word64

newtype RenderPass

VkRenderPass - Opaque handle to a render pass object

See Also

CommandBufferInheritanceInfo, FramebufferCreateInfo, GraphicsPipelineCreateInfo, RenderPassBeginInfo, createRenderPass, createRenderPass2, createRenderPass2KHR, destroyRenderPass, getRenderAreaGranularity

Constructors

RenderPass Word64

newtype PipelineCache

VkPipelineCache - Opaque handle to a pipeline cache object

See Also

createComputePipelines, createGraphicsPipelines, createPipelineCache, createRayTracingPipelinesNV, destroyPipelineCache, getPipelineCacheData, mergePipelineCaches

Constructors

PipelineCache Word64

newtype DescriptorUpdateTemplate

VkDescriptorUpdateTemplate - Opaque handle to a descriptor update template

See Also

cmdPushDescriptorSetWithTemplateKHR, createDescriptorUpdateTemplate, createDescriptorUpdateTemplateKHR, destroyDescriptorUpdateTemplate, destroyDescriptorUpdateTemplateKHR, updateDescriptorSetWithTemplate, updateDescriptorSetWithTemplateKHR

Constructors

DescriptorUpdateTemplate Word64

newtype SamplerYcbcrConversion

VkSamplerYcbcrConversion - Opaque handle to a device-specific sampler Y′CBCR conversion description

See Also

SamplerYcbcrConversionInfo, createSamplerYcbcrConversion, createSamplerYcbcrConversionKHR, destroySamplerYcbcrConversion, destroySamplerYcbcrConversionKHR

Constructors

SamplerYcbcrConversion Word64