Graphics.Vulkan.Extensions.VK_AMD_display_native_hdr

Documentation

withCStructDisplayNativeHdrSurfaceCapabilitiesAMD :: DisplayNativeHdrSurfaceCapabilitiesAMD -> (VkDisplayNativeHdrSurfaceCapabilitiesAMD -> IO a) -> IO a

A function to temporarily allocate memory for a VkDisplayNativeHdrSurfaceCapabilitiesAMD and marshal a DisplayNativeHdrSurfaceCapabilitiesAMD into it. The VkDisplayNativeHdrSurfaceCapabilitiesAMD is only valid inside the provided computation and must not be returned out of it.

fromCStructDisplayNativeHdrSurfaceCapabilitiesAMD :: VkDisplayNativeHdrSurfaceCapabilitiesAMD -> IO DisplayNativeHdrSurfaceCapabilitiesAMD

A function to read a VkDisplayNativeHdrSurfaceCapabilitiesAMD and all additional structures in the pointer chain into a DisplayNativeHdrSurfaceCapabilitiesAMD.

data DisplayNativeHdrSurfaceCapabilitiesAMD

VkDisplayNativeHdrSurfaceCapabilitiesAMD - Structure describing display native HDR specific capabilities of a surface

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

DisplayNativeHdrSurfaceCapabilitiesAMD

Fields

next :: Maybe SomeVkStruct

localDimmingSupport :: Bool

withCStructSwapchainDisplayNativeHdrCreateInfoAMD :: SwapchainDisplayNativeHdrCreateInfoAMD -> (VkSwapchainDisplayNativeHdrCreateInfoAMD -> IO a) -> IO a

A function to temporarily allocate memory for a VkSwapchainDisplayNativeHdrCreateInfoAMD and marshal a SwapchainDisplayNativeHdrCreateInfoAMD into it. The VkSwapchainDisplayNativeHdrCreateInfoAMD is only valid inside the provided computation and must not be returned out of it.

fromCStructSwapchainDisplayNativeHdrCreateInfoAMD :: VkSwapchainDisplayNativeHdrCreateInfoAMD -> IO SwapchainDisplayNativeHdrCreateInfoAMD

A function to read a VkSwapchainDisplayNativeHdrCreateInfoAMD and all additional structures in the pointer chain into a SwapchainDisplayNativeHdrCreateInfoAMD.

data SwapchainDisplayNativeHdrCreateInfoAMD

VkSwapchainDisplayNativeHdrCreateInfoAMD - Structure specifying display native HDR parameters of a newly created swapchain object

Description

If the pNext chain of VkSwapchainCreateInfoKHR does not contain this structure, the default value for localDimmingEnable is VK_TRUE, meaning local dimming is initially enabled for the swapchain.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_SWAPCHAIN_DISPLAY_NATIVE_HDR_CREATE_INFO_AMD

Valid Usage

It is only valid to set localDimmingEnable to VK_TRUE if VkDisplayNativeHdrSurfaceCapabilitiesAMD::localDimmingSupport is supported.

See Also

VkBool32, VkStructureType

Constructors

SwapchainDisplayNativeHdrCreateInfoAMD

Fields

next :: Maybe SomeVkStruct

localDimmingEnable :: Bool

setLocalDimmingAMD :: Device -> SwapchainKHR -> Bool -> IO ()

vkSetLocalDimmingAMD - Set Local Dimming

Parameters

device is the device associated with swapChain.

swapChain handle to enable local dimming.

localDimmingEnable specifies whether local dimming is enabled for the swapchain.

Valid Usage (Implicit)

device must be a valid VkDevice handle

swapChain must be a valid VkSwapchainKHR handle

Both of device, and swapChain must have been created, allocated, or retrieved from the same VkInstance

Valid Usage

It is only valid to call vkSetLocalDimmingAMD if VkDisplayNativeHdrSurfaceCapabilitiesAMD::localDimmingSupport is supported.

See Also

VkBool32, VkDevice, VkSwapchainKHR

pattern AMD_DISPLAY_NATIVE_HDR_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern AMD_DISPLAY_NATIVE_HDR_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_DISPLAY_NATIVE_HDR_SURFACE_CAPABILITIES_AMD :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SWAPCHAIN_DISPLAY_NATIVE_HDR_CREATE_INFO_AMD :: a ~ StructureType => a

pattern COLOR_SPACE_DISPLAY_NATIVE_AMD :: a ~ ColorSpaceKHR => a