Graphics.Vulkan.Extensions.VK_AMD_memory_overallocation_behavior

Documentation

data DeviceMemoryOverallocationCreateInfoAMD

VkDeviceMemoryOverallocationCreateInfoAMD - Specify memory overallocation behavior for a Vulkan device

Valid Usage (Implicit)

See Also

MemoryOverallocationBehaviorAMD, StructureType

Constructors

DeviceMemoryOverallocationCreateInfoAMD

Fields

overallocationBehavior :: MemoryOverallocationBehaviorAMD

overallocationBehavior must be a valid MemoryOverallocationBehaviorAMD value

newtype MemoryOverallocationBehaviorAMD

VkMemoryOverallocationBehaviorAMD - Specify memory overallocation behavior

See Also

DeviceMemoryOverallocationCreateInfoAMD

Constructors

MemoryOverallocationBehaviorAMD Int32

Bundled Patterns

pattern MEMORY_OVERALLOCATION_BEHAVIOR_DEFAULT_AMD :: MemoryOverallocationBehaviorAMD
MEMORY_OVERALLOCATION_BEHAVIOR_DEFAULT_AMD lets the implementation decide if overallocation should be allowed.

pattern MEMORY_OVERALLOCATION_BEHAVIOR_ALLOWED_AMD :: MemoryOverallocationBehaviorAMD
MEMORY_OVERALLOCATION_BEHAVIOR_ALLOWED_AMD specifies overallocation is allowed if platform permits.

pattern MEMORY_OVERALLOCATION_BEHAVIOR_DISALLOWED_AMD :: MemoryOverallocationBehaviorAMD
MEMORY_OVERALLOCATION_BEHAVIOR_DISALLOWED_AMD specifies the application is not allowed to allocate device memory beyond the heap sizes reported by PhysicalDeviceMemoryProperties. Allocations that are not explicitly made by the application within the scope of the Vulkan instance are not accounted for.

type AMD_MEMORY_OVERALLOCATION_BEHAVIOR_SPEC_VERSION = 1

pattern AMD_MEMORY_OVERALLOCATION_BEHAVIOR_SPEC_VERSION :: forall a. Integral a => a

type AMD_MEMORY_OVERALLOCATION_BEHAVIOR_EXTENSION_NAME = "VK_AMD_memory_overallocation_behavior"

pattern AMD_MEMORY_OVERALLOCATION_BEHAVIOR_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a