Graphics.Vulkan.Extensions.VK_AMD_rasterization_order

Documentation

withCStructPipelineRasterizationStateRasterizationOrderAMD :: PipelineRasterizationStateRasterizationOrderAMD -> (VkPipelineRasterizationStateRasterizationOrderAMD -> IO a) -> IO a

A function to temporarily allocate memory for a VkPipelineRasterizationStateRasterizationOrderAMD and marshal a PipelineRasterizationStateRasterizationOrderAMD into it. The VkPipelineRasterizationStateRasterizationOrderAMD is only valid inside the provided computation and must not be returned out of it.

fromCStructPipelineRasterizationStateRasterizationOrderAMD :: VkPipelineRasterizationStateRasterizationOrderAMD -> IO PipelineRasterizationStateRasterizationOrderAMD

A function to read a VkPipelineRasterizationStateRasterizationOrderAMD and all additional structures in the pointer chain into a PipelineRasterizationStateRasterizationOrderAMD.

data PipelineRasterizationStateRasterizationOrderAMD

VkPipelineRasterizationStateRasterizationOrderAMD - Structure defining rasterization order for a graphics pipeline

Valid Usage (Implicit)

If the https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#VK_AMD_rasterization_order device extension is not enabled or the application does not request a particular rasterization order through specifying a VkPipelineRasterizationStateRasterizationOrderAMD structure then the rasterization order used by the graphics pipeline defaults to VK_RASTERIZATION_ORDER_STRICT_AMD.

See Also

VkRasterizationOrderAMD, VkStructureType

Constructors

PipelineRasterizationStateRasterizationOrderAMD

Fields

next :: Maybe SomeVkStruct

rasterizationOrder :: RasterizationOrderAMD

type RasterizationOrderAMD = VkRasterizationOrderAMD

VkRasterizationOrderAMD - Specify rasterization order for a graphics pipeline

See Also

VkPipelineRasterizationStateRasterizationOrderAMD

pattern RASTERIZATION_ORDER_STRICT_AMD :: a ~ RasterizationOrderAMD => a

VK_RASTERIZATION_ORDER_STRICT_AMD specifies that operations for each primitive in a subpass must occur in primitive order.

pattern RASTERIZATION_ORDER_RELAXED_AMD :: a ~ RasterizationOrderAMD => a

VK_RASTERIZATION_ORDER_RELAXED_AMD specifies that operations for each primitive in a subpass may not occur in primitive order.

pattern AMD_RASTERIZATION_ORDER_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern AMD_RASTERIZATION_ORDER_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_RASTERIZATION_ORDER_AMD :: a ~ StructureType => a