Graphics.Vulkan.Extensions.VK_AMD_shader_core_properties

Documentation

withCStructPhysicalDeviceShaderCorePropertiesAMD :: PhysicalDeviceShaderCorePropertiesAMD -> (VkPhysicalDeviceShaderCorePropertiesAMD -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceShaderCorePropertiesAMD and marshal a PhysicalDeviceShaderCorePropertiesAMD into it. The VkPhysicalDeviceShaderCorePropertiesAMD is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceShaderCorePropertiesAMD :: VkPhysicalDeviceShaderCorePropertiesAMD -> IO PhysicalDeviceShaderCorePropertiesAMD

A function to read a VkPhysicalDeviceShaderCorePropertiesAMD and all additional structures in the pointer chain into a PhysicalDeviceShaderCorePropertiesAMD.

data PhysicalDeviceShaderCorePropertiesAMD

VkPhysicalDeviceShaderCorePropertiesAMD - Structure describing shader core properties that can be supported by an implementation

Members

The members of the VkPhysicalDeviceShaderCorePropertiesAMD structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceShaderCorePropertiesAMD structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

PhysicalDeviceShaderCorePropertiesAMD

Fields

next :: Maybe SomeVkStruct

shaderEngineCount :: Word32

shaderArraysPerEngineCount :: Word32

computeUnitsPerShaderArray :: Word32

simdPerComputeUnit :: Word32

wavefrontsPerSimd :: Word32

wavefrontSize :: Word32

sgprsPerSimd :: Word32

minSgprAllocation :: Word32

maxSgprAllocation :: Word32

sgprAllocationGranularity :: Word32

vgprsPerSimd :: Word32

minVgprAllocation :: Word32

maxVgprAllocation :: Word32

vgprAllocationGranularity :: Word32

pattern AMD_SHADER_CORE_PROPERTIES_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern AMD_SHADER_CORE_PROPERTIES_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_CORE_PROPERTIES_AMD :: a ~ StructureType => a