Graphics.Vulkan.Extensions.VK_AMD_shader_info

Documentation

type ShaderInfoTypeAMD = VkShaderInfoTypeAMD

VkShaderInfoTypeAMD - Enum specifying which type of shader info to query

See Also

vkGetShaderInfoAMD

pattern SHADER_INFO_TYPE_STATISTICS_AMD :: a ~ ShaderInfoTypeAMD => a

VK_SHADER_INFO_TYPE_STATISTICS_AMD specifies that device resources used by a shader will be queried.

pattern SHADER_INFO_TYPE_BINARY_AMD :: a ~ ShaderInfoTypeAMD => a

VK_SHADER_INFO_TYPE_BINARY_AMD specifies that implementation-specific information will be queried.

pattern SHADER_INFO_TYPE_DISASSEMBLY_AMD :: a ~ ShaderInfoTypeAMD => a

VK_SHADER_INFO_TYPE_DISASSEMBLY_AMD specifies that human-readable dissassembly of a shader.

withCStructShaderResourceUsageAMD :: ShaderResourceUsageAMD -> (VkShaderResourceUsageAMD -> IO a) -> IO a

A function to temporarily allocate memory for a VkShaderResourceUsageAMD and marshal a ShaderResourceUsageAMD into it. The VkShaderResourceUsageAMD is only valid inside the provided computation and must not be returned out of it.

fromCStructShaderResourceUsageAMD :: VkShaderResourceUsageAMD -> IO ShaderResourceUsageAMD

A function to read a VkShaderResourceUsageAMD and all additional structures in the pointer chain into a ShaderResourceUsageAMD.

data ShaderResourceUsageAMD

VkShaderResourceUsageAMD - Resource usage information about a particular shader within a pipeline

See Also

VkShaderStatisticsInfoAMD

Constructors

ShaderResourceUsageAMD

Fields

numUsedVgprs :: Word32

numUsedSgprs :: Word32

ldsSizePerLocalWorkGroup :: Word32

ldsUsageSizeInBytes :: CSize

scratchMemUsageInBytes :: CSize

withCStructShaderStatisticsInfoAMD :: ShaderStatisticsInfoAMD -> (VkShaderStatisticsInfoAMD -> IO a) -> IO a

A function to temporarily allocate memory for a VkShaderStatisticsInfoAMD and marshal a ShaderStatisticsInfoAMD into it. The VkShaderStatisticsInfoAMD is only valid inside the provided computation and must not be returned out of it.

fromCStructShaderStatisticsInfoAMD :: VkShaderStatisticsInfoAMD -> IO ShaderStatisticsInfoAMD

A function to read a VkShaderStatisticsInfoAMD and all additional structures in the pointer chain into a ShaderStatisticsInfoAMD.

data ShaderStatisticsInfoAMD

VkShaderStatisticsInfoAMD - Statistical information about a particular shader within a pipeline

Description

Some implementations may merge multiple logical shader stages together in a single shader. In such cases, shaderStageMask will contain a bitmask of all of the stages that are active within that shader. Consequently, if specifying those stages as input to vkGetShaderInfoAMD, the same output information may be returned for all such shader stage queries.

The number of available VGPRs and SGPRs (numAvailableVgprs and numAvailableSgprs respectively) are the shader-addressable subset of physical registers that is given as a limit to the compiler for register assignment. These values may further be limited by implementations due to performance optimizations where register pressure is a bottleneck.

See Also

VkShaderResourceUsageAMD, VkShaderStageFlags

Constructors

ShaderStatisticsInfoAMD

Fields

shaderStageMask :: ShaderStageFlags

resourceUsage :: ShaderResourceUsageAMD

numPhysicalVgprs :: Word32

numPhysicalSgprs :: Word32

numAvailableVgprs :: Word32

numAvailableSgprs :: Word32

computeWorkGroupSize :: (Word32, Word32, Word32)

getNumShaderInfoAMD :: Device -> Pipeline -> ShaderStageFlagBits -> ShaderInfoTypeAMD -> IO (VkResult, CSize)

vkGetShaderInfoAMD - Get information about a shader in a pipeline

Parameters

device is the device that created pipeline.

pipeline is the target of the query.

shaderStage identifies the particular shader within the pipeline about which information is being queried.

infoType describes what kind of information is being queried.

pInfoSize is a pointer to a value related to the amount of data the query returns, as described below.

pInfo is either NULL or a pointer to a buffer.

Description

If pInfo is NULL, then the maximum size of the information that can be retrieved about the shader, in bytes, is returned in pInfoSize. Otherwise, pInfoSize must point to a variable set by the user to the size of the buffer, in bytes, pointed to by pInfo, and on return the variable is overwritten with the amount of data actually written to pInfo.

If pInfoSize is less than the maximum size that can be retrieved by the pipeline cache, then at most pInfoSize bytes will be written to pInfo, and vkGetShaderInfoAMD will return VK_INCOMPLETE.

Not all information is available for every shader and implementations may not support all kinds of information for any shader. When a certain type of information is unavailable, the function returns VK_ERROR_FEATURE_NOT_PRESENT.

If information is successfully and fully queried, the function will return VK_SUCCESS.

For infoType VK_SHADER_INFO_TYPE_STATISTICS_AMD, an instance of VkShaderStatisticsInfoAMD will be written to the buffer pointed to by pInfo. This structure will be populated with statistics regarding the physical device resources used by that shader along with other miscellaneous information and is described in further detail below.

For infoType VK_SHADER_INFO_TYPE_DISASSEMBLY_AMD, pInfo points to a UTF-8 null-terminated string containing human-readable disassembly. The exact formatting and contents of the disassembly string are vendor-specific.

The formatting and contents of all other types of information, including infoType VK_SHADER_INFO_TYPE_BINARY_AMD, are left to the vendor and are not further specified by this extension.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pipeline must be a valid VkPipeline handle

shaderStage must be a valid VkShaderStageFlagBits value

infoType must be a valid VkShaderInfoTypeAMD value

pInfoSize must be a valid pointer to a size_t value

If the value referenced by pInfoSize is not 0, and pInfo is not NULL, pInfo must be a valid pointer to an array of pInfoSize bytes

pipeline must have been created, allocated, or retrieved from device

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_FEATURE_NOT_PRESENT

VK_ERROR_OUT_OF_HOST_MEMORY

See Also

VkDevice, VkPipeline, VkShaderInfoTypeAMD, VkShaderStageFlagBits

getShaderInfoAMD :: Device -> Pipeline -> ShaderStageFlagBits -> ShaderInfoTypeAMD -> CSize -> IO (VkResult, ByteString)

vkGetShaderInfoAMD - Get information about a shader in a pipeline

Parameters

device is the device that created pipeline.

pipeline is the target of the query.

shaderStage identifies the particular shader within the pipeline about which information is being queried.

infoType describes what kind of information is being queried.

pInfoSize is a pointer to a value related to the amount of data the query returns, as described below.

pInfo is either NULL or a pointer to a buffer.

Description

If pInfo is NULL, then the maximum size of the information that can be retrieved about the shader, in bytes, is returned in pInfoSize. Otherwise, pInfoSize must point to a variable set by the user to the size of the buffer, in bytes, pointed to by pInfo, and on return the variable is overwritten with the amount of data actually written to pInfo.

If pInfoSize is less than the maximum size that can be retrieved by the pipeline cache, then at most pInfoSize bytes will be written to pInfo, and vkGetShaderInfoAMD will return VK_INCOMPLETE.

Not all information is available for every shader and implementations may not support all kinds of information for any shader. When a certain type of information is unavailable, the function returns VK_ERROR_FEATURE_NOT_PRESENT.

If information is successfully and fully queried, the function will return VK_SUCCESS.

For infoType VK_SHADER_INFO_TYPE_STATISTICS_AMD, an instance of VkShaderStatisticsInfoAMD will be written to the buffer pointed to by pInfo. This structure will be populated with statistics regarding the physical device resources used by that shader along with other miscellaneous information and is described in further detail below.

For infoType VK_SHADER_INFO_TYPE_DISASSEMBLY_AMD, pInfo points to a UTF-8 null-terminated string containing human-readable disassembly. The exact formatting and contents of the disassembly string are vendor-specific.

The formatting and contents of all other types of information, including infoType VK_SHADER_INFO_TYPE_BINARY_AMD, are left to the vendor and are not further specified by this extension.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pipeline must be a valid VkPipeline handle

shaderStage must be a valid VkShaderStageFlagBits value

infoType must be a valid VkShaderInfoTypeAMD value

pInfoSize must be a valid pointer to a size_t value

If the value referenced by pInfoSize is not 0, and pInfo is not NULL, pInfo must be a valid pointer to an array of pInfoSize bytes

pipeline must have been created, allocated, or retrieved from device

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_FEATURE_NOT_PRESENT

VK_ERROR_OUT_OF_HOST_MEMORY

See Also

VkDevice, VkPipeline, VkShaderInfoTypeAMD, VkShaderStageFlagBits

getAllShaderInfoAMD :: Device -> Pipeline -> ShaderStageFlagBits -> ShaderInfoTypeAMD -> IO ByteString

Returns all the values available from getShaderInfoAMD.

pattern AMD_SHADER_INFO_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern AMD_SHADER_INFO_SPEC_VERSION :: Integral a => a