Graphics.Vulkan.Extensions.VK_ANDROID_external_memory_android_hardware_buffer

Documentation

withCStructAndroidHardwareBufferFormatPropertiesANDROID :: AndroidHardwareBufferFormatPropertiesANDROID -> (VkAndroidHardwareBufferFormatPropertiesANDROID -> IO a) -> IO a

A function to temporarily allocate memory for a VkAndroidHardwareBufferFormatPropertiesANDROID and marshal a AndroidHardwareBufferFormatPropertiesANDROID into it. The VkAndroidHardwareBufferFormatPropertiesANDROID is only valid inside the provided computation and must not be returned out of it.

fromCStructAndroidHardwareBufferFormatPropertiesANDROID :: VkAndroidHardwareBufferFormatPropertiesANDROID -> IO AndroidHardwareBufferFormatPropertiesANDROID

A function to read a VkAndroidHardwareBufferFormatPropertiesANDROID and all additional structures in the pointer chain into a AndroidHardwareBufferFormatPropertiesANDROID.

data AndroidHardwareBufferFormatPropertiesANDROID

VkAndroidHardwareBufferFormatPropertiesANDROID - Structure describing the image format properties of an Android hardware buffer

Description

If the Android hardware buffer has one of the formats listed in the Format Equivalence table, then format must have the equivalent Vulkan format listed in the table. Otherwise, format may be VK_FORMAT_UNDEFINED, indicating the Android hardware buffer can only be used with an external format.

The formatFeatures member must include VK_FORMAT_FEATURE_SAMPLED_IMAGE_BIT and at least one of VK_FORMAT_FEATURE_MIDPOINT_CHROMA_SAMPLES_BIT or VK_FORMAT_FEATURE_COSITED_CHROMA_SAMPLES_BIT, and should include VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_LINEAR_BIT and VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_LINEAR_FILTER_BIT.

Note

The formatFeatures member only indicates the features available when using an external-format image created from the Android hardware buffer. Images from Android hardware buffers with a format other than VK_FORMAT_UNDEFINED are subject to the format capabilities obtained from vkGetPhysicalDeviceFormatProperties2, and vkGetPhysicalDeviceImageFormatProperties2 with appropriate parameters. These sets of features are independent of each other, e.g. the external format will support sampler Y’CBCR conversion even if the non-external format does not, and writing to non-external format images is possible but writing to external format images is not.

Android hardware buffers with the same external format must have the same support for VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_LINEAR_BIT, VK_FORMAT_FEATURE_MIDPOINT_CHROMA_SAMPLES_BIT, VK_FORMAT_FEATURE_COSITED_CHROMA_SAMPLES_BIT, VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_LINEAR_FILTER_BIT, VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_SEPARATE_RECONSTRUCTION_FILTER_BIT, and VK_FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_FORCEABLE_BIT. in formatFeatures. Other format features may differ between Android hardware buffers that have the same external format. This allows applications to use the same VkSamplerYcbcrConversion object (and samplers and pipelines created from them) for any Android hardware buffers that have the same external format.

If format is not VK_FORMAT_UNDEFINED, then the value of samplerYcbcrConversionComponents must be valid when used as the components member of VkSamplerYcbcrConversionCreateInfo with that format. If format is VK_FORMAT_UNDEFINED, all members of samplerYcbcrConversionComponents must be VK_COMPONENT_SWIZZLE_IDENTITY.

Implementations may not always be able to determine the color model, numerical range, or chroma offsets of the image contents, so the values in VkAndroidHardwareBufferFormatPropertiesANDROID are only suggestions. Applications should treat these values as sensible defaults to use in the absence of more reliable information obtained through some other means. If the underlying physical device is also usable via OpenGL ES with the GL_OES_EGL_image_external extension, the implementation should suggest values that will produce similar sampled values as would be obtained by sampling the same external image via samplerExternalOES in OpenGL ES using equivalent sampler parameters.

Note

Since GL_OES_EGL_image_external does not require the same sampling and conversion calculations as Vulkan does, achieving identical results between APIs may not be possible on some implementations.

Valid Usage (Implicit)

See Also

VkChromaLocation, VkComponentMapping, VkFormat, VkFormatFeatureFlags, VkSamplerYcbcrModelConversion, VkSamplerYcbcrRange, VkStructureType

Constructors

AndroidHardwareBufferFormatPropertiesANDROID

Fields

next :: Maybe SomeVkStruct

format :: Format

externalFormat :: Word64

formatFeatures :: FormatFeatureFlags

samplerYcbcrConversionComponents :: ComponentMapping

suggestedYcbcrModel :: SamplerYcbcrModelConversion

suggestedYcbcrRange :: SamplerYcbcrRange

suggestedXChromaOffset :: ChromaLocation

suggestedYChromaOffset :: ChromaLocation

withCStructAndroidHardwareBufferPropertiesANDROID :: AndroidHardwareBufferPropertiesANDROID -> (VkAndroidHardwareBufferPropertiesANDROID -> IO a) -> IO a

A function to temporarily allocate memory for a VkAndroidHardwareBufferPropertiesANDROID and marshal a AndroidHardwareBufferPropertiesANDROID into it. The VkAndroidHardwareBufferPropertiesANDROID is only valid inside the provided computation and must not be returned out of it.

fromCStructAndroidHardwareBufferPropertiesANDROID :: VkAndroidHardwareBufferPropertiesANDROID -> IO AndroidHardwareBufferPropertiesANDROID

A function to read a VkAndroidHardwareBufferPropertiesANDROID and all additional structures in the pointer chain into a AndroidHardwareBufferPropertiesANDROID.

data AndroidHardwareBufferPropertiesANDROID

VkAndroidHardwareBufferPropertiesANDROID - Properties of External Memory Android Hardware Buffers

Valid Usage (Implicit)

See Also

VkDeviceSize, VkStructureType, vkGetAndroidHardwareBufferPropertiesANDROID

Constructors

AndroidHardwareBufferPropertiesANDROID

Fields

next :: Maybe SomeVkStruct

allocationSize :: DeviceSize

memoryTypeBits :: Word32

withCStructAndroidHardwareBufferUsageANDROID :: AndroidHardwareBufferUsageANDROID -> (VkAndroidHardwareBufferUsageANDROID -> IO a) -> IO a

A function to temporarily allocate memory for a VkAndroidHardwareBufferUsageANDROID and marshal a AndroidHardwareBufferUsageANDROID into it. The VkAndroidHardwareBufferUsageANDROID is only valid inside the provided computation and must not be returned out of it.

fromCStructAndroidHardwareBufferUsageANDROID :: VkAndroidHardwareBufferUsageANDROID -> IO AndroidHardwareBufferUsageANDROID

A function to read a VkAndroidHardwareBufferUsageANDROID and all additional structures in the pointer chain into a AndroidHardwareBufferUsageANDROID.

data AndroidHardwareBufferUsageANDROID

VkAndroidHardwareBufferUsageANDROID - Struct containing Android hardware buffer usage flags

Description

The androidHardwareBufferUsage field must include Android hardware buffer usage flags listed in the AHardwareBuffer Usage Equivalence table when the corresponding Vulkan image usage or image creation flags are included in the usage or flags fields of VkPhysicalDeviceImageFormatInfo2. It must include at least one GPU usage flag (AHARDWAREBUFFER_USAGE_GPU_*), even if none of the corresponding Vulkan usages or flags are requested.

Note

Requiring at least one GPU usage flag ensures that Android hardware buffer memory will be allocated in a memory pool accessible to the Vulkan implementation, and that specializing the memory layout based on usage flags does not prevent it from being compatible with Vulkan. Implementations may avoid unnecessary restrictions caused by this requirement by using vendor usage flags to indicate that only the Vulkan uses indicated in VkImageFormatProperties2 are required.

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

AndroidHardwareBufferUsageANDROID

Fields

next :: Maybe SomeVkStruct

androidHardwareBufferUsage :: Word64

withCStructExternalFormatANDROID :: ExternalFormatANDROID -> (VkExternalFormatANDROID -> IO a) -> IO a

A function to temporarily allocate memory for a VkExternalFormatANDROID and marshal a ExternalFormatANDROID into it. The VkExternalFormatANDROID is only valid inside the provided computation and must not be returned out of it.

fromCStructExternalFormatANDROID :: VkExternalFormatANDROID -> IO ExternalFormatANDROID

A function to read a VkExternalFormatANDROID and all additional structures in the pointer chain into a ExternalFormatANDROID.

data ExternalFormatANDROID

VkExternalFormatANDROID - Structure containing an Android hardware buffer external format

Description

If externalFormat is zero, the effect is as if the VkExternalFormatANDROID structure was not present. Otherwise, the image will have the specified external format.

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

ExternalFormatANDROID

Fields

next :: Maybe SomeVkStruct

externalFormat :: Word64

withCStructImportAndroidHardwareBufferInfoANDROID :: ImportAndroidHardwareBufferInfoANDROID -> (VkImportAndroidHardwareBufferInfoANDROID -> IO a) -> IO a

A function to temporarily allocate memory for a VkImportAndroidHardwareBufferInfoANDROID and marshal a ImportAndroidHardwareBufferInfoANDROID into it. The VkImportAndroidHardwareBufferInfoANDROID is only valid inside the provided computation and must not be returned out of it.

fromCStructImportAndroidHardwareBufferInfoANDROID :: VkImportAndroidHardwareBufferInfoANDROID -> IO ImportAndroidHardwareBufferInfoANDROID

A function to read a VkImportAndroidHardwareBufferInfoANDROID and all additional structures in the pointer chain into a ImportAndroidHardwareBufferInfoANDROID.

data ImportAndroidHardwareBufferInfoANDROID

VkImportAndroidHardwareBufferInfoANDROID - Import memory from an Android hardware buffer

Description

If the vkAllocateMemory command succeeds, the implementation must acquire a reference to the imported hardware buffer, which it must release when the device memory object is freed. If the command fails, the implementation must not retain a reference.

Valid Usage

If buffer is not NULL, Android hardware buffers must be supported for import, as reported by VkExternalImageFormatProperties or VkExternalBufferProperties.

If buffer is not NULL, it must be a valid Android hardware buffer object with AHardwareBuffer_Desc::format and AHardwareBuffer_Desc::usage compatible with Vulkan as described in Android Hardware Buffers.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMPORT_ANDROID_HARDWARE_BUFFER_INFO_ANDROID

buffer must be a valid pointer to an AHardwareBuffer value

See Also

VkStructureType

Constructors

ImportAndroidHardwareBufferInfoANDROID

Fields

next :: Maybe SomeVkStruct

buffer :: Ptr AHardwareBuffer

withCStructMemoryGetAndroidHardwareBufferInfoANDROID :: MemoryGetAndroidHardwareBufferInfoANDROID -> (VkMemoryGetAndroidHardwareBufferInfoANDROID -> IO a) -> IO a

A function to temporarily allocate memory for a VkMemoryGetAndroidHardwareBufferInfoANDROID and marshal a MemoryGetAndroidHardwareBufferInfoANDROID into it. The VkMemoryGetAndroidHardwareBufferInfoANDROID is only valid inside the provided computation and must not be returned out of it.

fromCStructMemoryGetAndroidHardwareBufferInfoANDROID :: VkMemoryGetAndroidHardwareBufferInfoANDROID -> IO MemoryGetAndroidHardwareBufferInfoANDROID

A function to read a VkMemoryGetAndroidHardwareBufferInfoANDROID and all additional structures in the pointer chain into a MemoryGetAndroidHardwareBufferInfoANDROID.

data MemoryGetAndroidHardwareBufferInfoANDROID

VkMemoryGetAndroidHardwareBufferInfoANDROID - Structure describing an Android hardware buffer memory export operation

Valid Usage

VK_EXTERNAL_MEMORY_HANDLE_TYPE_ANDROID_HARDWARE_BUFFER_BIT_ANDROID must have been included in VkExportMemoryAllocateInfoKHR::handleTypes when memory was created.

If the pNext chain of the VkMemoryAllocateInfo used to allocate memory included a VkMemoryDedicatedAllocateInfo with non-NULL image member, then that image must already be bound to memory.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_MEMORY_GET_ANDROID_HARDWARE_BUFFER_INFO_ANDROID

pNext must be NULL

memory must be a valid VkDeviceMemory handle

See Also

VkDeviceMemory, VkStructureType, vkGetMemoryAndroidHardwareBufferANDROID

Constructors

MemoryGetAndroidHardwareBufferInfoANDROID

Fields

next :: Maybe SomeVkStruct

memory :: DeviceMemory

getAndroidHardwareBufferPropertiesANDROID :: Device -> Ptr AHardwareBuffer -> IO AndroidHardwareBufferPropertiesANDROID

vkGetAndroidHardwareBufferPropertiesANDROID - Get Properties of External Memory Android Hardware Buffers

Parameters

device is the logical device that will be importing buffer.

buffer is the Android hardware buffer which will be imported.

pProperties is a pointer to a VkAndroidHardwareBufferPropertiesANDROID structure in which the properties of buffer are returned.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_INVALID_EXTERNAL_HANDLE_KHR

See Also

VkAndroidHardwareBufferPropertiesANDROID, VkDevice

getMemoryAndroidHardwareBufferANDROID :: Device -> MemoryGetAndroidHardwareBufferInfoANDROID -> IO (Ptr AHardwareBuffer)

vkGetMemoryAndroidHardwareBufferANDROID - Get an Android hardware buffer for a memory object

Parameters

device is the logical device that created the device memory being exported.

pInfo is a pointer to an instance of the VkMemoryGetAndroidHardwareBufferInfoANDROID structure containing parameters of the export operation.

pBuffer will return an Android hardware buffer representing the underlying resources of the device memory object.

Description

Each call to vkGetMemoryAndroidHardwareBufferANDROID must return an Android hardware buffer with a new reference acquired in addition to the reference held by the VkDeviceMemory. To avoid leaking resources, the application must release the reference by calling AHardwareBuffer_release when it is no longer needed. When called with the same handle in VkMemoryGetAndroidHardwareBufferInfoANDROID::memory, vkGetMemoryAndroidHardwareBufferANDROID must return the same Android hardware buffer object. If the device memory was created by importing an Android hardware buffer, vkGetMemoryAndroidHardwareBufferANDROID must return that same Android hardware buffer object.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_TOO_MANY_OBJECTS

VK_ERROR_OUT_OF_HOST_MEMORY

See Also

VkDevice, VkMemoryGetAndroidHardwareBufferInfoANDROID

pattern ANDROID_EXTERNAL_MEMORY_ANDROID_HARDWARE_BUFFER_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern ANDROID_EXTERNAL_MEMORY_ANDROID_HARDWARE_BUFFER_SPEC_VERSION :: Integral a => a

pattern EXTERNAL_MEMORY_HANDLE_TYPE_ANDROID_HARDWARE_BUFFER_BIT_ANDROID :: a ~ ExternalMemoryHandleTypeFlagBits => a

pattern STRUCTURE_TYPE_ANDROID_HARDWARE_BUFFER_USAGE_ANDROID :: a ~ StructureType => a

pattern STRUCTURE_TYPE_ANDROID_HARDWARE_BUFFER_PROPERTIES_ANDROID :: a ~ StructureType => a

pattern STRUCTURE_TYPE_ANDROID_HARDWARE_BUFFER_FORMAT_PROPERTIES_ANDROID :: a ~ StructureType => a

pattern STRUCTURE_TYPE_IMPORT_ANDROID_HARDWARE_BUFFER_INFO_ANDROID :: a ~ StructureType => a

pattern STRUCTURE_TYPE_MEMORY_GET_ANDROID_HARDWARE_BUFFER_INFO_ANDROID :: a ~ StructureType => a

pattern STRUCTURE_TYPE_EXTERNAL_FORMAT_ANDROID :: a ~ StructureType => a