Graphics.Vulkan.Extensions.VK_EXT_acquire_xlib_display

Documentation

acquireXlibDisplayEXT :: PhysicalDevice -> DisplayKHR -> IO Display

vkAcquireXlibDisplayEXT - Acquire access to a VkDisplayKHR using Xlib

Parameters

physicalDevice The physical device the display is on.

dpy A connection to the X11 server that currently owns display.

display The display the caller wishes to control in Vulkan.

Description

All permissions necessary to control the display are granted to the Vulkan instance associated with physicalDevice until the display is released or the X11 connection specified by dpy is terminated. Permission to access the display may be temporarily revoked during periods when the X11 server from which control was acquired itself looses access to display. During such periods, operations which require access to the display must fail with an approriate error code. If the X11 server associated with dpy does not own display, or if permission to access it has already been acquired by another entity, the call must return the error code VK_ERROR_INITIALIZATION_FAILED.

Note

One example of when an X11 server loses access to a display is when it loses ownership of its virtual terminal.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_INITIALIZATION_FAILED

See Also

VkDisplayKHR, VkPhysicalDevice

getRandROutputDisplayEXT :: PhysicalDevice -> RROutput -> IO (Display, DisplayKHR)

vkGetRandROutputDisplayEXT - Query the VkDisplayKHR corresponding to an X11 RandR Output

Parameters

physicalDevice The physical device to query the display handle on.

dpy A connection to the X11 server from which rrOutput was queried.

rrOutput An X11 RandR output ID.

pDisplay The corresponding VkDisplayKHR handle will be returned here.

Description

If there is no VkDisplayKHR corresponding to rrOutput on physicalDevice, VK_NULL_HANDLE must be returned in pDisplay.

Return Codes

Success

- VK_SUCCESS

See Also

VkDisplayKHR, VkPhysicalDevice

pattern EXT_ACQUIRE_XLIB_DISPLAY_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_ACQUIRE_XLIB_DISPLAY_SPEC_VERSION :: Integral a => a