Graphics.Vulkan.Extensions.VK_EXT_astc_decode_mode

Documentation

withCStructImageViewASTCDecodeModeEXT :: ImageViewASTCDecodeModeEXT -> (VkImageViewASTCDecodeModeEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkImageViewASTCDecodeModeEXT and marshal a ImageViewASTCDecodeModeEXT into it. The VkImageViewASTCDecodeModeEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructImageViewASTCDecodeModeEXT :: VkImageViewASTCDecodeModeEXT -> IO ImageViewASTCDecodeModeEXT

A function to read a VkImageViewASTCDecodeModeEXT and all additional structures in the pointer chain into a ImageViewASTCDecodeModeEXT.

data ImageViewASTCDecodeModeEXT

VkImageViewASTCDecodeModeEXT - Structure describing the ASTC decode mode for an image view

Valid Usage

decodeMode must be one of VK_FORMAT_R16G16B16A16_SFLOAT, VK_FORMAT_R8G8B8A8_UNORM, or VK_FORMAT_E5B9G9R9_UFLOAT_PACK32

If the decodeModeSharedExponent feature is not enabled, decodeMode must not be VK_FORMAT_E5B9G9R9_UFLOAT_PACK32

If decodeMode is VK_FORMAT_R8G8B8A8_UNORM the image view must not include blocks using any of the ASTC HDR modes

format of the image view must be one of VK_FORMAT_ASTC_4x4_UNORM_BLOCK, VK_FORMAT_ASTC_4x4_SRGB_BLOCK, VK_FORMAT_ASTC_5x4_UNORM_BLOCK, VK_FORMAT_ASTC_5x4_SRGB_BLOCK, VK_FORMAT_ASTC_5x5_UNORM_BLOCK, VK_FORMAT_ASTC_5x5_SRGB_BLOCK, VK_FORMAT_ASTC_6x5_UNORM_BLOCK, VK_FORMAT_ASTC_6x5_SRGB_BLOCK, VK_FORMAT_ASTC_6x6_UNORM_BLOCK, VK_FORMAT_ASTC_6x6_SRGB_BLOCK, VK_FORMAT_ASTC_8x5_UNORM_BLOCK, VK_FORMAT_ASTC_8x5_SRGB_BLOCK, VK_FORMAT_ASTC_8x6_UNORM_BLOCK, VK_FORMAT_ASTC_8x6_SRGB_BLOCK, VK_FORMAT_ASTC_8x8_UNORM_BLOCK, VK_FORMAT_ASTC_8x8_SRGB_BLOCK, VK_FORMAT_ASTC_10x5_UNORM_BLOCK, VK_FORMAT_ASTC_10x5_SRGB_BLOCK, VK_FORMAT_ASTC_10x6_UNORM_BLOCK, VK_FORMAT_ASTC_10x6_SRGB_BLOCK, VK_FORMAT_ASTC_10x8_UNORM_BLOCK, VK_FORMAT_ASTC_10x8_SRGB_BLOCK, VK_FORMAT_ASTC_10x10_UNORM_BLOCK, VK_FORMAT_ASTC_10x10_SRGB_BLOCK, VK_FORMAT_ASTC_12x10_UNORM_BLOCK, VK_FORMAT_ASTC_12x10_SRGB_BLOCK, VK_FORMAT_ASTC_12x12_UNORM_BLOCK, or VK_FORMAT_ASTC_12x12_SRGB_BLOCK

If format uses sRGB encoding then the decodeMode has no effect.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMAGE_VIEW_ASTC_DECODE_MODE_EXT

decodeMode must be a valid VkFormat value

See Also

VkFormat, VkStructureType

Constructors

ImageViewASTCDecodeModeEXT

Fields

next :: Maybe SomeVkStruct

decodeMode :: Format

withCStructPhysicalDeviceASTCDecodeFeaturesEXT :: PhysicalDeviceASTCDecodeFeaturesEXT -> (VkPhysicalDeviceASTCDecodeFeaturesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceASTCDecodeFeaturesEXT and marshal a PhysicalDeviceASTCDecodeFeaturesEXT into it. The VkPhysicalDeviceASTCDecodeFeaturesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceASTCDecodeFeaturesEXT :: VkPhysicalDeviceASTCDecodeFeaturesEXT -> IO PhysicalDeviceASTCDecodeFeaturesEXT

A function to read a VkPhysicalDeviceASTCDecodeFeaturesEXT and all additional structures in the pointer chain into a PhysicalDeviceASTCDecodeFeaturesEXT.

data PhysicalDeviceASTCDecodeFeaturesEXT

VkPhysicalDeviceASTCDecodeFeaturesEXT - Structure describing ASTC decode mode features

Members

The members of the VkPhysicalDeviceASTCDecodeFeaturesEXT structure describe the following features:

Description

If the VkPhysicalDeviceASTCDecodeFeaturesEXT structure is included in the pNext chain of vkGetPhysicalDeviceFeatures2KHR, it is filled with values indicating whether each feature is supported. VkPhysicalDeviceASTCDecodeFeaturesEXT can also be used in the pNext chain of vkCreateDevice to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceASTCDecodeFeaturesEXT

Fields

next :: Maybe SomeVkStruct

decodeModeSharedExponent :: Bool

pattern EXT_ASTC_DECODE_MODE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_ASTC_DECODE_MODE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_IMAGE_VIEW_ASTC_DECODE_MODE_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_ASTC_DECODE_FEATURES_EXT :: a ~ StructureType => a