Graphics.Vulkan.Extensions.VK_EXT_blend_operation_advanced

Documentation

type BlendOverlapEXT = VkBlendOverlapEXT

VkBlendOverlapEXT - Enumerant specifying the blend overlap parameter

Description

'

+-----------------------------------+-----------------------------------+
| Overlap Mode                      | Weighting Equations               |
+===================================+===================================+
| 'Graphics.Vulkan.C.Extensions.VK_ | \[                                |
| EXT_blend_operation_advanced.VK_B |                \begin{aligned}    |
| LEND_OVERLAP_UNCORRELATED_EXT'    |                                   |
|                                   |                p_0(A_s,A_d) & = A |
|                                   | _sA_d \\                          |
|                                   |                                   |
|                                   |                p_1(A_s,A_d) & = A |
|                                   | _s(1-A_d) \\                      |
|                                   |                                   |
|                                   |                p_2(A_s,A_d) & = A |
|                                   | _d(1-A_s) \\                      |
|                                   |                                   |
|                                   |              \end{aligned}\]      |
+-----------------------------------+-----------------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_ | \[                                |
| EXT_blend_operation_advanced.VK_B |                \begin{aligned}    |
| LEND_OVERLAP_CONJOINT_EXT'        |                                   |
|                                   |                p_0(A_s,A_d) & = m |
|                                   | in(A_s,A_d) \\                    |
|                                   |                                   |
|                                   |                p_1(A_s,A_d) & = m |
|                                   | ax(A_s-A_d,0) \\                  |
|                                   |                                   |
|                                   |                p_2(A_s,A_d) & = m |
|                                   | ax(A_d-A_s,0) \\                  |
|                                   |                                   |
|                                   |              \end{aligned}\]      |
+-----------------------------------+-----------------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_ | \[                                |
| EXT_blend_operation_advanced.VK_B |                \begin{aligned}    |
| LEND_OVERLAP_DISJOINT_EXT'        |                                   |
|                                   |                p_0(A_s,A_d) & = m |
|                                   | ax(A_s+A_d-1,0) \\                |
|                                   |                                   |
|                                   |                p_1(A_s,A_d) & = m |
|                                   | in(A_s,1-A_d) \\                  |
|                                   |                                   |
|                                   |                p_2(A_s,A_d) & = m |
|                                   | in(A_d,1-A_s) \\                  |
|                                   |                                   |
|                                   |              \end{aligned}\]      |
+-----------------------------------+-----------------------------------+

Advanced Blend Overlap Modes

See Also

VkPipelineColorBlendAdvancedStateCreateInfoEXT

pattern BLEND_OVERLAP_UNCORRELATED_EXT :: a ~ BlendOverlapEXT => a

VK_BLEND_OVERLAP_UNCORRELATED_EXT specifies that there is no correlation between the source and destination coverage.

pattern BLEND_OVERLAP_DISJOINT_EXT :: a ~ BlendOverlapEXT => a

VK_BLEND_OVERLAP_DISJOINT_EXT specifies that the source and destination coverage are considered to have minimal overlap.

pattern BLEND_OVERLAP_CONJOINT_EXT :: a ~ BlendOverlapEXT => a

VK_BLEND_OVERLAP_CONJOINT_EXT specifies that the source and destination coverage are considered to have maximal overlap.

withCStructPhysicalDeviceBlendOperationAdvancedFeaturesEXT :: PhysicalDeviceBlendOperationAdvancedFeaturesEXT -> (VkPhysicalDeviceBlendOperationAdvancedFeaturesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceBlendOperationAdvancedFeaturesEXT and marshal a PhysicalDeviceBlendOperationAdvancedFeaturesEXT into it. The VkPhysicalDeviceBlendOperationAdvancedFeaturesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceBlendOperationAdvancedFeaturesEXT :: VkPhysicalDeviceBlendOperationAdvancedFeaturesEXT -> IO PhysicalDeviceBlendOperationAdvancedFeaturesEXT

A function to read a VkPhysicalDeviceBlendOperationAdvancedFeaturesEXT and all additional structures in the pointer chain into a PhysicalDeviceBlendOperationAdvancedFeaturesEXT.

data PhysicalDeviceBlendOperationAdvancedFeaturesEXT

VkPhysicalDeviceBlendOperationAdvancedFeaturesEXT - Structure describing advanced blending features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceBlendOperationAdvancedFeaturesEXT structure describe the following features:

Description

If the VkPhysicalDeviceBlendOperationAdvancedFeaturesEXT structure is included in the pNext chain of VkPhysicalDeviceFeatures2, it is filled with values indicating whether each feature is supported. VkPhysicalDeviceBlendOperationAdvancedFeaturesEXT can also be used in pNext chain of VkDeviceCreateInfo to enable the features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceBlendOperationAdvancedFeaturesEXT

Fields

next :: Maybe SomeVkStruct

advancedBlendCoherentOperations :: Bool

withCStructPhysicalDeviceBlendOperationAdvancedPropertiesEXT :: PhysicalDeviceBlendOperationAdvancedPropertiesEXT -> (VkPhysicalDeviceBlendOperationAdvancedPropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceBlendOperationAdvancedPropertiesEXT and marshal a PhysicalDeviceBlendOperationAdvancedPropertiesEXT into it. The VkPhysicalDeviceBlendOperationAdvancedPropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceBlendOperationAdvancedPropertiesEXT :: VkPhysicalDeviceBlendOperationAdvancedPropertiesEXT -> IO PhysicalDeviceBlendOperationAdvancedPropertiesEXT

A function to read a VkPhysicalDeviceBlendOperationAdvancedPropertiesEXT and all additional structures in the pointer chain into a PhysicalDeviceBlendOperationAdvancedPropertiesEXT.

data PhysicalDeviceBlendOperationAdvancedPropertiesEXT

VkPhysicalDeviceBlendOperationAdvancedPropertiesEXT - Structure describing advanced blending limits that can be supported by an implementation

Members

The members of the VkPhysicalDeviceBlendOperationAdvancedPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceBlendOperationAdvancedPropertiesEXT structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceBlendOperationAdvancedPropertiesEXT

Fields

next :: Maybe SomeVkStruct

advancedBlendMaxColorAttachments :: Word32

advancedBlendIndependentBlend :: Bool

advancedBlendNonPremultipliedSrcColor :: Bool

advancedBlendNonPremultipliedDstColor :: Bool

advancedBlendCorrelatedOverlap :: Bool

advancedBlendAllOperations :: Bool

withCStructPipelineColorBlendAdvancedStateCreateInfoEXT :: PipelineColorBlendAdvancedStateCreateInfoEXT -> (VkPipelineColorBlendAdvancedStateCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPipelineColorBlendAdvancedStateCreateInfoEXT and marshal a PipelineColorBlendAdvancedStateCreateInfoEXT into it. The VkPipelineColorBlendAdvancedStateCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPipelineColorBlendAdvancedStateCreateInfoEXT :: VkPipelineColorBlendAdvancedStateCreateInfoEXT -> IO PipelineColorBlendAdvancedStateCreateInfoEXT

A function to read a VkPipelineColorBlendAdvancedStateCreateInfoEXT and all additional structures in the pointer chain into a PipelineColorBlendAdvancedStateCreateInfoEXT.

data PipelineColorBlendAdvancedStateCreateInfoEXT

VkPipelineColorBlendAdvancedStateCreateInfoEXT - Structure specifying parameters that affect advanced blend operations

Description

If this structure is not present, srcPremultiplied and dstPremultiplied are both considered to be VK_TRUE, and blendOverlap is considered to be VK_BLEND_OVERLAP_UNCORRELATED_EXT.

Valid Usage

If the non-premultiplied source color property is not supported, srcPremultiplied must be VK_TRUE

If the non-premultiplied destination color property is not supported, dstPremultiplied must be VK_TRUE

If the correlated overlap property is not supported, blendOverlap must be VK_BLEND_OVERLAP_UNCORRELATED_EXT

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_ADVANCED_STATE_CREATE_INFO_EXT

blendOverlap must be a valid VkBlendOverlapEXT value

See Also

VkBlendOverlapEXT, VkBool32, VkStructureType

Constructors

PipelineColorBlendAdvancedStateCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

srcPremultiplied :: Bool

dstPremultiplied :: Bool

blendOverlap :: BlendOverlapEXT

pattern EXT_BLEND_OPERATION_ADVANCED_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_BLEND_OPERATION_ADVANCED_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_BLEND_OPERATION_ADVANCED_FEATURES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_BLEND_OPERATION_ADVANCED_PROPERTIES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_ADVANCED_STATE_CREATE_INFO_EXT :: a ~ StructureType => a

pattern BLEND_OP_ZERO_EXT :: a ~ BlendOp => a

pattern BLEND_OP_SRC_EXT :: a ~ BlendOp => a

pattern BLEND_OP_DST_EXT :: a ~ BlendOp => a

pattern BLEND_OP_SRC_OVER_EXT :: a ~ BlendOp => a

pattern BLEND_OP_DST_OVER_EXT :: a ~ BlendOp => a

pattern BLEND_OP_SRC_IN_EXT :: a ~ BlendOp => a

pattern BLEND_OP_DST_IN_EXT :: a ~ BlendOp => a

pattern BLEND_OP_SRC_OUT_EXT :: a ~ BlendOp => a

pattern BLEND_OP_DST_OUT_EXT :: a ~ BlendOp => a

pattern BLEND_OP_SRC_ATOP_EXT :: a ~ BlendOp => a

pattern BLEND_OP_DST_ATOP_EXT :: a ~ BlendOp => a

pattern BLEND_OP_XOR_EXT :: a ~ BlendOp => a

pattern BLEND_OP_MULTIPLY_EXT :: a ~ BlendOp => a

pattern BLEND_OP_SCREEN_EXT :: a ~ BlendOp => a

pattern BLEND_OP_OVERLAY_EXT :: a ~ BlendOp => a

pattern BLEND_OP_DARKEN_EXT :: a ~ BlendOp => a

pattern BLEND_OP_LIGHTEN_EXT :: a ~ BlendOp => a

pattern BLEND_OP_COLORDODGE_EXT :: a ~ BlendOp => a

pattern BLEND_OP_COLORBURN_EXT :: a ~ BlendOp => a

pattern BLEND_OP_HARDLIGHT_EXT :: a ~ BlendOp => a

pattern BLEND_OP_SOFTLIGHT_EXT :: a ~ BlendOp => a

pattern BLEND_OP_DIFFERENCE_EXT :: a ~ BlendOp => a

pattern BLEND_OP_EXCLUSION_EXT :: a ~ BlendOp => a

pattern BLEND_OP_INVERT_EXT :: a ~ BlendOp => a

pattern BLEND_OP_INVERT_RGB_EXT :: a ~ BlendOp => a

pattern BLEND_OP_LINEARDODGE_EXT :: a ~ BlendOp => a

pattern BLEND_OP_LINEARBURN_EXT :: a ~ BlendOp => a

pattern BLEND_OP_VIVIDLIGHT_EXT :: a ~ BlendOp => a

pattern BLEND_OP_LINEARLIGHT_EXT :: a ~ BlendOp => a

pattern BLEND_OP_PINLIGHT_EXT :: a ~ BlendOp => a

pattern BLEND_OP_HARDMIX_EXT :: a ~ BlendOp => a

pattern BLEND_OP_HSL_HUE_EXT :: a ~ BlendOp => a

pattern BLEND_OP_HSL_SATURATION_EXT :: a ~ BlendOp => a

pattern BLEND_OP_HSL_COLOR_EXT :: a ~ BlendOp => a

pattern BLEND_OP_HSL_LUMINOSITY_EXT :: a ~ BlendOp => a

pattern BLEND_OP_PLUS_EXT :: a ~ BlendOp => a

pattern BLEND_OP_PLUS_CLAMPED_EXT :: a ~ BlendOp => a

pattern BLEND_OP_PLUS_CLAMPED_ALPHA_EXT :: a ~ BlendOp => a

pattern BLEND_OP_PLUS_DARKER_EXT :: a ~ BlendOp => a

pattern BLEND_OP_MINUS_EXT :: a ~ BlendOp => a

pattern BLEND_OP_MINUS_CLAMPED_EXT :: a ~ BlendOp => a

pattern BLEND_OP_CONTRAST_EXT :: a ~ BlendOp => a

pattern BLEND_OP_INVERT_OVG_EXT :: a ~ BlendOp => a

pattern BLEND_OP_RED_EXT :: a ~ BlendOp => a

pattern BLEND_OP_GREEN_EXT :: a ~ BlendOp => a

pattern BLEND_OP_BLUE_EXT :: a ~ BlendOp => a

pattern ACCESS_COLOR_ATTACHMENT_READ_NONCOHERENT_BIT_EXT :: a ~ AccessFlagBits => a