Graphics.Vulkan.Extensions.VK_EXT_buffer_device_address

Documentation

withCStructBufferDeviceAddressCreateInfoEXT :: BufferDeviceAddressCreateInfoEXT -> (VkBufferDeviceAddressCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkBufferDeviceAddressCreateInfoEXT and marshal a BufferDeviceAddressCreateInfoEXT into it. The VkBufferDeviceAddressCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructBufferDeviceAddressCreateInfoEXT :: VkBufferDeviceAddressCreateInfoEXT -> IO BufferDeviceAddressCreateInfoEXT

A function to read a VkBufferDeviceAddressCreateInfoEXT and all additional structures in the pointer chain into a BufferDeviceAddressCreateInfoEXT.

data BufferDeviceAddressCreateInfoEXT

VkBufferDeviceAddressCreateInfoEXT - Request a specific address for a buffer

Description

If deviceAddress is zero, no specific address is requested.

If deviceAddress is not zero, deviceAddress must be an address retrieved from an identically created buffer on the same implementation. The buffer must also be bound to an identically created VkDeviceMemory object.

If this structure is not present, it is as if deviceAddress is zero.

Apps should avoid creating buffers with app-provided addresses and implementation-provided addresses in the same process, to reduce the likelihood of VK_ERROR_INVALID_DEVICE_ADDRESS_EXT errors.

Note

The expected usage for this is that a trace capture/replay tool will add the VK_BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT_EXT flag to all buffers that use VK_BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT_EXT, and during capture will save the queried device addresses in the trace. During replay, the buffers will be created specifying the original address so any address values stored in the trace data will remain valid.

Implementations are expected to separate such buffers in the GPU address space so normal allocations will avoid using these addresses. Apps/tools should avoid mixing app-provided and implementation-provided addresses for buffers created with VK_BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT_EXT, to avoid address space allocation conflicts.

Valid Usage (Implicit)

See Also

VkDeviceAddress, VkStructureType

Constructors

BufferDeviceAddressCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

deviceAddress :: DeviceAddress

withCStructBufferDeviceAddressInfoEXT :: BufferDeviceAddressInfoEXT -> (VkBufferDeviceAddressInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkBufferDeviceAddressInfoEXT and marshal a BufferDeviceAddressInfoEXT into it. The VkBufferDeviceAddressInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructBufferDeviceAddressInfoEXT :: VkBufferDeviceAddressInfoEXT -> IO BufferDeviceAddressInfoEXT

A function to read a VkBufferDeviceAddressInfoEXT and all additional structures in the pointer chain into a BufferDeviceAddressInfoEXT.

data BufferDeviceAddressInfoEXT

VkBufferDeviceAddressInfoEXT - Structure specifying the buffer to query an address for

Valid Usage

If buffer is non-sparse and was not created with the VK_BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT_EXT flag, then it must be bound completely and contiguously to a single VkDeviceMemory object

buffer must have been created with VK_BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT_EXT

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_BUFFER_DEVICE_ADDRESS_INFO_EXT

pNext must be NULL

buffer must be a valid VkBuffer handle

See Also

VkBuffer, VkStructureType, vkGetBufferDeviceAddressEXT

Constructors

BufferDeviceAddressInfoEXT

Fields

next :: Maybe SomeVkStruct

buffer :: Buffer

type DeviceAddress = VkDeviceAddress

VkDeviceAddress - Vulkan device address type

See Also

VkBufferDeviceAddressCreateInfoEXT

type PhysicalDeviceBufferAddressFeaturesEXT = PhysicalDeviceBufferDeviceAddressFeaturesEXT

withCStructPhysicalDeviceBufferDeviceAddressFeaturesEXT :: PhysicalDeviceBufferDeviceAddressFeaturesEXT -> (VkPhysicalDeviceBufferDeviceAddressFeaturesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceBufferDeviceAddressFeaturesEXT and marshal a PhysicalDeviceBufferDeviceAddressFeaturesEXT into it. The VkPhysicalDeviceBufferDeviceAddressFeaturesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceBufferDeviceAddressFeaturesEXT :: VkPhysicalDeviceBufferDeviceAddressFeaturesEXT -> IO PhysicalDeviceBufferDeviceAddressFeaturesEXT

A function to read a VkPhysicalDeviceBufferDeviceAddressFeaturesEXT and all additional structures in the pointer chain into a PhysicalDeviceBufferDeviceAddressFeaturesEXT.

data PhysicalDeviceBufferDeviceAddressFeaturesEXT

VkPhysicalDeviceBufferDeviceAddressFeaturesEXT - Structure describing buffer address features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceBufferDeviceAddressFeaturesEXT structure describe the following features:

Description

Note

bufferDeviceAddressMultiDevice exists to allow certain legacy platforms to be able to support bufferDeviceAddress without needing to support shared GPU virtual addresses for multi-device configurations.

See vkGetBufferDeviceAddressEXT for more information.

If the VkPhysicalDeviceBufferDeviceAddressFeaturesEXT structure is included in the pNext chain of VkPhysicalDeviceFeatures2KHR, it is filled with values indicating whether the feature is supported. VkPhysicalDeviceBufferDeviceAddressFeaturesEXT can also be used in the pNext chain of VkDeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceBufferDeviceAddressFeaturesEXT

Fields

next :: Maybe SomeVkStruct

bufferDeviceAddress :: Bool

bufferDeviceAddressCaptureReplay :: Bool

bufferDeviceAddressMultiDevice :: Bool

getBufferDeviceAddressEXT :: Device -> BufferDeviceAddressInfoEXT -> IO VkDeviceAddress

vkGetBufferDeviceAddressEXT - Query an address of a buffer

Parameters

device is the logical device that the buffer was created on.

pInfo is a pointer to an instance of the VkBufferDeviceAddressInfoEXT structure specifying the buffer to retrieve an address for.

Description

The 64-bit return value is an address of the start of pInfo::buffer. The address range starting at this value and whose size is the size of the buffer can be used in a shader to access the memory bound to that buffer, using the SPV_EXT_physical_storage_buffer extension and the PhysicalStorageBufferEXT storage class. For example, this value can be stored in a uniform buffer, and the shader can read the value from the uniform buffer and use it to do a dependent read/write to this buffer. A value of zero is reserved as a “null” pointer and must not be returned as a valid buffer device address. All loads, stores, and atomics in a shader through PhysicalStorageBufferEXT pointers must access addresses in the address range of some buffer.

If the buffer was created with a non-zero value of VkBufferDeviceAddressCreateInfoEXT::deviceAddress, the return value will be the same address.

Valid Usage

The bufferDeviceAddress feature must be enabled

If device was created with multiple physical devices, then the bufferDeviceAddressMultiDevice feature must be enabled

Valid Usage (Implicit)

device must be a valid VkDevice handle

pInfo must be a valid pointer to a valid VkBufferDeviceAddressInfoEXT structure

See Also

VkBufferDeviceAddressInfoEXT, VkDevice

pattern EXT_BUFFER_DEVICE_ADDRESS_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_BUFFER_DEVICE_ADDRESS_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_BUFFER_ADDRESS_FEATURES_EXT :: VkStructureType

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_BUFFER_DEVICE_ADDRESS_FEATURES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_BUFFER_DEVICE_ADDRESS_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_BUFFER_DEVICE_ADDRESS_CREATE_INFO_EXT :: a ~ StructureType => a

pattern BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT_EXT :: a ~ BufferUsageFlagBits => a

pattern BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT_EXT :: a ~ BufferCreateFlagBits => a

pattern ERROR_INVALID_DEVICE_ADDRESS_EXT :: a ~ Result => a