Graphics.Vulkan.Extensions.VK_EXT_conditional_rendering

Documentation

withCStructCommandBufferInheritanceConditionalRenderingInfoEXT :: CommandBufferInheritanceConditionalRenderingInfoEXT -> (VkCommandBufferInheritanceConditionalRenderingInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkCommandBufferInheritanceConditionalRenderingInfoEXT and marshal a CommandBufferInheritanceConditionalRenderingInfoEXT into it. The VkCommandBufferInheritanceConditionalRenderingInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructCommandBufferInheritanceConditionalRenderingInfoEXT :: VkCommandBufferInheritanceConditionalRenderingInfoEXT -> IO CommandBufferInheritanceConditionalRenderingInfoEXT

A function to read a VkCommandBufferInheritanceConditionalRenderingInfoEXT and all additional structures in the pointer chain into a CommandBufferInheritanceConditionalRenderingInfoEXT.

data CommandBufferInheritanceConditionalRenderingInfoEXT

VkCommandBufferInheritanceConditionalRenderingInfoEXT - Structure specifying command buffer inheritance info

Description

If this structure is not present, the behavior is as if conditionalRenderingEnable is VK_FALSE.

Valid Usage

If the inherited conditional rendering feature is not enabled, conditionalRenderingEnable must be VK_FALSE

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_COMMAND_BUFFER_INHERITANCE_CONDITIONAL_RENDERING_INFO_EXT

See Also

VkBool32, VkStructureType

Constructors

CommandBufferInheritanceConditionalRenderingInfoEXT

Fields

next :: Maybe SomeVkStruct

conditionalRenderingEnable :: Bool

withCStructConditionalRenderingBeginInfoEXT :: ConditionalRenderingBeginInfoEXT -> (VkConditionalRenderingBeginInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkConditionalRenderingBeginInfoEXT and marshal a ConditionalRenderingBeginInfoEXT into it. The VkConditionalRenderingBeginInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructConditionalRenderingBeginInfoEXT :: VkConditionalRenderingBeginInfoEXT -> IO ConditionalRenderingBeginInfoEXT

A function to read a VkConditionalRenderingBeginInfoEXT and all additional structures in the pointer chain into a ConditionalRenderingBeginInfoEXT.

data ConditionalRenderingBeginInfoEXT

VkConditionalRenderingBeginInfoEXT - Structure specifying conditional rendering begin info

Description

If the 32-bit value at offset in buffer memory is zero, then the rendering commands are discarded, otherwise they are executed as normal. If the value of the predicate in buffer memory changes while conditional rendering is active, the rendering commands may be discarded in an implementation-dependent way. Some implementations may latch the value of the predicate upon beginning conditional rendering while others may read it before every rendering command.

Valid Usage

If buffer is non-sparse then it must be bound completely and contiguously to a single VkDeviceMemory object

buffer must have been created with the VK_BUFFER_USAGE_CONDITIONAL_RENDERING_BIT_EXT bit set

offset must be less than the size of buffer by at least 32 bits.

offset must be a multiple of 4

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_CONDITIONAL_RENDERING_BEGIN_INFO_EXT

pNext must be NULL

buffer must be a valid VkBuffer handle

flags must be a valid combination of VkConditionalRenderingFlagBitsEXT values

See Also

VkBuffer, VkConditionalRenderingFlagsEXT, VkDeviceSize, VkStructureType, vkCmdBeginConditionalRenderingEXT

Constructors

ConditionalRenderingBeginInfoEXT

Fields

next :: Maybe SomeVkStruct

buffer :: Buffer

offset :: DeviceSize

flags :: ConditionalRenderingFlagsEXT

type ConditionalRenderingFlagBitsEXT = VkConditionalRenderingFlagBitsEXT

VkConditionalRenderingFlagBitsEXT - Specify the behavior of conditional rendering

See Also

VkConditionalRenderingFlagsEXT

pattern CONDITIONAL_RENDERING_INVERTED_BIT_EXT :: a ~ ConditionalRenderingFlagBitsEXT => a

VK_CONDITIONAL_RENDERING_INVERTED_BIT_EXT specifies the condition used to determine whether to discard rendering commands or not. That is, if the 32-bit predicate read from buffer memory at offset is zero, the rendering commands are not discarded, and if non zero, then they are discarded.

type ConditionalRenderingFlagsEXT = ConditionalRenderingFlagBitsEXT

VkConditionalRenderingFlagsEXT - Bitmask of VkConditionalRenderingFlagBitsEXT

Description

VkConditionalRenderingFlagsEXT is a bitmask type for setting a mask of zero or more VkConditionalRenderingFlagBitsEXT.

See Also

VkConditionalRenderingBeginInfoEXT, VkConditionalRenderingFlagBitsEXT

withCStructPhysicalDeviceConditionalRenderingFeaturesEXT :: PhysicalDeviceConditionalRenderingFeaturesEXT -> (VkPhysicalDeviceConditionalRenderingFeaturesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceConditionalRenderingFeaturesEXT and marshal a PhysicalDeviceConditionalRenderingFeaturesEXT into it. The VkPhysicalDeviceConditionalRenderingFeaturesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceConditionalRenderingFeaturesEXT :: VkPhysicalDeviceConditionalRenderingFeaturesEXT -> IO PhysicalDeviceConditionalRenderingFeaturesEXT

A function to read a VkPhysicalDeviceConditionalRenderingFeaturesEXT and all additional structures in the pointer chain into a PhysicalDeviceConditionalRenderingFeaturesEXT.

data PhysicalDeviceConditionalRenderingFeaturesEXT

VkPhysicalDeviceConditionalRenderingFeaturesEXT - Structure describing if a secondary command buffer can be executed if conditional rendering is active in the primary command buffer

Description

If the VkPhysicalDeviceConditionalRenderingFeaturesEXT structure is included in the pNext chain of VkPhysicalDeviceFeatures2, it is filled with values indicating the implementation-dependent behavior. VkPhysicalDeviceConditionalRenderingFeaturesEXT can also be used in pNext chain of VkDeviceCreateInfo to enable the features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceConditionalRenderingFeaturesEXT

Fields

next :: Maybe SomeVkStruct

conditionalRendering :: Bool

inheritedConditionalRendering :: Bool

cmdBeginConditionalRenderingEXT :: CommandBuffer -> ConditionalRenderingBeginInfoEXT -> IO ()

vkCmdBeginConditionalRenderingEXT - Define the beginning of a conditional rendering block

Parameters

commandBuffer is the command buffer into which this command will be recorded.

pConditionalRenderingBegin is a pointer to an instance of the VkConditionalRenderingBeginInfoEXT structure specifying the parameters of conditional rendering.

Valid Usage

Conditional rendering must not already be active

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pConditionalRenderingBegin must be a valid pointer to a valid VkConditionalRenderingBeginInfoEXT structure

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkConditionalRenderingBeginInfoEXT

cmdEndConditionalRenderingEXT :: CommandBuffer -> IO ()

vkCmdEndConditionalRenderingEXT - Define the end of a conditional rendering block

Parameters

commandBuffer is the command buffer into which this command will be recorded.

Description

Once ended, conditional rendering becomes inactive.

Valid Usage

Conditional rendering must be active

If conditional rendering was made active outside of a render pass instance, it must not be ended inside a render pass instance

If conditional rendering was made active within a subpass it must be ended in the same subpass

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer

pattern EXT_CONDITIONAL_RENDERING_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_CONDITIONAL_RENDERING_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_COMMAND_BUFFER_INHERITANCE_CONDITIONAL_RENDERING_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_CONDITIONAL_RENDERING_FEATURES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_CONDITIONAL_RENDERING_BEGIN_INFO_EXT :: a ~ StructureType => a

pattern ACCESS_CONDITIONAL_RENDERING_READ_BIT_EXT :: a ~ AccessFlagBits => a

pattern BUFFER_USAGE_CONDITIONAL_RENDERING_BIT_EXT :: a ~ BufferUsageFlagBits => a

pattern PIPELINE_STAGE_CONDITIONAL_RENDERING_BIT_EXT :: a ~ PipelineStageFlagBits => a