Graphics.Vulkan.Extensions.VK_EXT_conservative_rasterization

Documentation

type ConservativeRasterizationModeEXT = VkConservativeRasterizationModeEXT

VkConservativeRasterizationModeEXT - Specify the conservative rasterization mode

See Also

VkPipelineRasterizationConservativeStateCreateInfoEXT

pattern CONSERVATIVE_RASTERIZATION_MODE_DISABLED_EXT :: a ~ ConservativeRasterizationModeEXT => a

VK_CONSERVATIVE_RASTERIZATION_MODE_DISABLED_EXT specifies that conservative rasterization is disabled and rasterization proceeds as normal.

pattern CONSERVATIVE_RASTERIZATION_MODE_OVERESTIMATE_EXT :: a ~ ConservativeRasterizationModeEXT => a

VK_CONSERVATIVE_RASTERIZATION_MODE_OVERESTIMATE_EXT specifies that conservative rasterization is enabled in overestimation mode.

pattern CONSERVATIVE_RASTERIZATION_MODE_UNDERESTIMATE_EXT :: a ~ ConservativeRasterizationModeEXT => a

VK_CONSERVATIVE_RASTERIZATION_MODE_UNDERESTIMATE_EXT specifies that conservative rasterization is enabled in underestimation mode.

withCStructPhysicalDeviceConservativeRasterizationPropertiesEXT :: PhysicalDeviceConservativeRasterizationPropertiesEXT -> (VkPhysicalDeviceConservativeRasterizationPropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceConservativeRasterizationPropertiesEXT and marshal a PhysicalDeviceConservativeRasterizationPropertiesEXT into it. The VkPhysicalDeviceConservativeRasterizationPropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceConservativeRasterizationPropertiesEXT :: VkPhysicalDeviceConservativeRasterizationPropertiesEXT -> IO PhysicalDeviceConservativeRasterizationPropertiesEXT

A function to read a VkPhysicalDeviceConservativeRasterizationPropertiesEXT and all additional structures in the pointer chain into a PhysicalDeviceConservativeRasterizationPropertiesEXT.

data PhysicalDeviceConservativeRasterizationPropertiesEXT

VkPhysicalDeviceConservativeRasterizationPropertiesEXT - Structure describing conservative raster properties that can be supported by an implementation

Members

The members of the VkPhysicalDeviceConservativeRasterizationPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceConservativeRasterizationPropertiesEXT structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits and properties.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceConservativeRasterizationPropertiesEXT

Fields

next :: Maybe SomeVkStruct

primitiveOverestimationSize :: CFloat

maxExtraPrimitiveOverestimationSize :: CFloat

extraPrimitiveOverestimationSizeGranularity :: CFloat

primitiveUnderestimation :: Bool

conservativePointAndLineRasterization :: Bool

degenerateTrianglesRasterized :: Bool

degenerateLinesRasterized :: Bool

fullyCoveredFragmentShaderInputVariable :: Bool

conservativeRasterizationPostDepthCoverage :: Bool

type PipelineRasterizationConservativeStateCreateFlagsEXT = VkPipelineRasterizationConservativeStateCreateFlagsEXT

VkPipelineRasterizationConservativeStateCreateFlagsEXT - Reserved for future use

Description

VkPipelineRasterizationConservativeStateCreateFlagsEXT is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkPipelineRasterizationConservativeStateCreateInfoEXT

withCStructPipelineRasterizationConservativeStateCreateInfoEXT :: PipelineRasterizationConservativeStateCreateInfoEXT -> (VkPipelineRasterizationConservativeStateCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPipelineRasterizationConservativeStateCreateInfoEXT and marshal a PipelineRasterizationConservativeStateCreateInfoEXT into it. The VkPipelineRasterizationConservativeStateCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPipelineRasterizationConservativeStateCreateInfoEXT :: VkPipelineRasterizationConservativeStateCreateInfoEXT -> IO PipelineRasterizationConservativeStateCreateInfoEXT

A function to read a VkPipelineRasterizationConservativeStateCreateInfoEXT and all additional structures in the pointer chain into a PipelineRasterizationConservativeStateCreateInfoEXT.

data PipelineRasterizationConservativeStateCreateInfoEXT

VkPipelineRasterizationConservativeStateCreateInfoEXT - Structure specifying conservative raster state

Valid Usage (Implicit)

See Also

VkConservativeRasterizationModeEXT, VkPipelineRasterizationConservativeStateCreateFlagsEXT, VkStructureType

Constructors

PipelineRasterizationConservativeStateCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

flags :: PipelineRasterizationConservativeStateCreateFlagsEXT

conservativeRasterizationMode :: ConservativeRasterizationModeEXT

extraPrimitiveOverestimationSize :: CFloat

pattern EXT_CONSERVATIVE_RASTERIZATION_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_CONSERVATIVE_RASTERIZATION_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_CONSERVATIVE_RASTERIZATION_PROPERTIES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PIPELINE_RASTERIZATION_CONSERVATIVE_STATE_CREATE_INFO_EXT :: a ~ StructureType => a