Graphics.Vulkan.Extensions.VK_EXT_debug_report

Documentation

withCStructDebugReportCallbackCreateInfoEXT :: DebugReportCallbackCreateInfoEXT -> (VkDebugReportCallbackCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDebugReportCallbackCreateInfoEXT and marshal a DebugReportCallbackCreateInfoEXT into it. The VkDebugReportCallbackCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDebugReportCallbackCreateInfoEXT :: VkDebugReportCallbackCreateInfoEXT -> IO DebugReportCallbackCreateInfoEXT

A function to read a VkDebugReportCallbackCreateInfoEXT and all additional structures in the pointer chain into a DebugReportCallbackCreateInfoEXT.

data DebugReportCallbackCreateInfoEXT

VkDebugReportCallbackCreateInfoEXT - Structure specifying parameters of a newly created debug report callback

Description

For each VkDebugReportCallbackEXT that is created the VkDebugReportCallbackCreateInfoEXT::flags determine when that VkDebugReportCallbackCreateInfoEXT::pfnCallback is called. When an event happens, the implementation will do a bitwise AND of the event’s VkDebugReportFlagBitsEXT flags to each VkDebugReportCallbackEXT object’s flags. For each non-zero result the corresponding callback will be called. The callback will come directly from the component that detected the event, unless some other layer intercepts the calls for its own purposes (filter them in a different way, log to a system error log, etc.).

An application may receive multiple callbacks if multiple VkDebugReportCallbackEXT objects were created. A callback will always be executed in the same thread as the originating Vulkan call.

A callback may be called from multiple threads simultaneously (if the application is making Vulkan calls from multiple threads).

Valid Usage (Implicit)

See Also

PFN_vkDebugReportCallbackEXT, VkDebugReportFlagsEXT, VkStructureType, vkCreateDebugReportCallbackEXT

Constructors

DebugReportCallbackCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

flags :: DebugReportFlagsEXT

pfnCallback :: PFN_vkDebugReportCallbackEXT

userData :: Ptr ()

type DebugReportCallbackEXT = VkDebugReportCallbackEXT

VkDebugReportCallbackEXT - Opaque handle to a debug report callback object

See Also

vkCreateDebugReportCallbackEXT, vkDestroyDebugReportCallbackEXT

type DebugReportFlagBitsEXT = VkDebugReportFlagBitsEXT

VkDebugReportFlagBitsEXT - Bitmask specifying events which cause a debug report callback

See Also

VkDebugReportFlagsEXT

pattern DEBUG_REPORT_INFORMATION_BIT_EXT :: a ~ DebugReportFlagBitsEXT => a

VK_DEBUG_REPORT_INFORMATION_BIT_EXT specifies an informational message such as resource details that may be handy when debugging an application.

pattern DEBUG_REPORT_WARNING_BIT_EXT :: a ~ DebugReportFlagBitsEXT => a

VK_DEBUG_REPORT_WARNING_BIT_EXT specifies use of Vulkan that may expose an app bug. Such cases may not be immediately harmful, such as a fragment shader outputting to a location with no attachment. Other cases may point to behavior that is almost certainly bad when unintended such as using an image whose memory has not been filled. In general if you see a warning but you know that the behavior is intended/desired, then simply ignore the warning.

pattern DEBUG_REPORT_PERFORMANCE_WARNING_BIT_EXT :: a ~ DebugReportFlagBitsEXT => a

VK_DEBUG_REPORT_PERFORMANCE_WARNING_BIT_EXT specifies a potentially non-optimal use of Vulkan, e.g. using vkCmdClearColorImage when setting VkAttachmentDescription::loadOp to VK_ATTACHMENT_LOAD_OP_CLEAR would have worked.

pattern DEBUG_REPORT_ERROR_BIT_EXT :: a ~ DebugReportFlagBitsEXT => a

VK_DEBUG_REPORT_ERROR_BIT_EXT specifies that the application has violated a valid usage condition of the specification.

pattern DEBUG_REPORT_DEBUG_BIT_EXT :: a ~ DebugReportFlagBitsEXT => a

VK_DEBUG_REPORT_DEBUG_BIT_EXT specifies diagnostic information from the implementation and layers.

type DebugReportFlagsEXT = DebugReportFlagBitsEXT

VkDebugReportFlagsEXT - Bitmask of VkDebugReportFlagBitsEXT

Description

VkDebugReportFlagsEXT is a bitmask type for setting a mask of zero or more VkDebugReportFlagBitsEXT.

See Also

VkDebugReportCallbackCreateInfoEXT, VkDebugReportFlagBitsEXT, vkDebugReportMessageEXT

type DebugReportObjectTypeEXT = VkDebugReportObjectTypeEXT

VkDebugReportObjectTypeEXT - Specify the type of an object handle

Description

'

+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | Vulkan Handle Type        |
| ug_report.VkDebugReportObjectTypeEXT'    |                           |
+==========================================+===========================+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | Unknown\/Undefined Handle |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_UN |                           |
| KNOWN_EXT'                               |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_IN | .DeviceInitialization.VkI |
| STANCE_EXT'                              | nstance'                  |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_PH | .DeviceInitialization.VkP |
| YSICAL_DEVICE_EXT'                       | hysicalDevice'            |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_DE | .DeviceInitialization.VkD |
| VICE_EXT'                                | evice'                    |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_QU | .Queue.VkQueue'           |
| EUE_EXT'                                 |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_SE | .Queue.VkSemaphore'       |
| MAPHORE_EXT'                             |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_CO | .Queue.VkCommandBuffer'   |
| MMAND_BUFFER_EXT'                        |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_FE | .Queue.VkFence'           |
| NCE_EXT'                                 |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_DE | .Memory.VkDeviceMemory'   |
| VICE_MEMORY_EXT'                         |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_BU | .MemoryManagement.VkBuffe |
| FFER_EXT'                                | r'                        |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_IM | .MemoryManagement.VkImage |
| AGE_EXT'                                 | '                         |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_EV | .Event.VkEvent'           |
| ENT_EXT'                                 |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_QU | .Query.VkQueryPool'       |
| ERY_POOL_EXT'                            |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_BU | .BufferView.VkBufferView' |
| FFER_VIEW_EXT'                           |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_IM | .ImageView.VkImageView'   |
| AGE_VIEW_EXT'                            |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_SH | .Shader.VkShaderModule'   |
| ADER_MODULE_EXT'                         |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_PI | .PipelineCache.VkPipeline |
| PELINE_CACHE_EXT'                        | Cache'                    |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_PI | .Pipeline.VkPipelineLayou |
| PELINE_LAYOUT_EXT'                       | t'                        |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_RE | .Pipeline.VkRenderPass'   |
| NDER_PASS_EXT'                           |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_PI | .Pipeline.VkPipeline'     |
| PELINE_EXT'                              |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_DE | .PipelineLayout.VkDescrip |
| SCRIPTOR_SET_LAYOUT_EXT'                 | torSetLayout'             |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_SA | .Sampler.VkSampler'       |
| MPLER_EXT'                               |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_DE | .DescriptorSet.VkDescript |
| SCRIPTOR_POOL_EXT'                       | orPool'                   |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_DE | .DescriptorSet.VkDescript |
| SCRIPTOR_SET_EXT'                        | orSet'                    |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_FR | .Pass.VkFramebuffer'      |
| AMEBUFFER_EXT'                           |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_CO | .CommandPool.VkCommandPoo |
| MMAND_POOL_EXT'                          | l'                        |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Extens |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_DE | ions.VK_EXT_debug_report. |
| BUG_REPORT_CALLBACK_EXT_EXT'             | VkDebugReportCallbackEXT' |
+------------------------------------------+---------------------------+

VkDebugReportObjectTypeEXT and Vulkan Handle Relationship

Note

The primary expected use of VK_ERROR_VALIDATION_FAILED_EXT is for validation layer testing. It is not expected that an application would see this error code during normal use of the validation layers.

See Also

VkDebugMarkerObjectNameInfoEXT, VkDebugMarkerObjectTagInfoEXT, vkDebugReportMessageEXT

pattern DEBUG_REPORT_OBJECT_TYPE_UNKNOWN_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_INSTANCE_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_PHYSICAL_DEVICE_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_DEVICE_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_QUEUE_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_SEMAPHORE_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_COMMAND_BUFFER_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_FENCE_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_DEVICE_MEMORY_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_BUFFER_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_IMAGE_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_EVENT_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_QUERY_POOL_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_BUFFER_VIEW_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_IMAGE_VIEW_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_SHADER_MODULE_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_PIPELINE_CACHE_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_PIPELINE_LAYOUT_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_RENDER_PASS_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_PIPELINE_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_LAYOUT_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_SAMPLER_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_POOL_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_FRAMEBUFFER_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_COMMAND_POOL_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_SURFACE_KHR_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_SWAPCHAIN_KHR_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_DEBUG_REPORT_CALLBACK_EXT_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_DISPLAY_KHR_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_DISPLAY_MODE_KHR_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_OBJECT_TABLE_NVX_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_INDIRECT_COMMANDS_LAYOUT_NVX_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_VALIDATION_CACHE_EXT_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern DEBUG_REPORT_OBJECT_TYPE_ACCELERATION_STRUCTURE_NV_EXT :: a ~ DebugReportObjectTypeEXT => a

createDebugReportCallbackEXT :: Instance -> DebugReportCallbackCreateInfoEXT -> Maybe AllocationCallbacks -> IO DebugReportCallbackEXT

vkCreateDebugReportCallbackEXT - Create a debug report callback object

Parameters

instance the instance the callback will be logged on.

pCreateInfo points to a VkDebugReportCallbackCreateInfoEXT structure which defines the conditions under which this callback will be called.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pCallback is a pointer to record the VkDebugReportCallbackEXT object created.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkDebugReportCallbackCreateInfoEXT structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pCallback must be a valid pointer to a VkDebugReportCallbackEXT handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

See Also

VkAllocationCallbacks, VkDebugReportCallbackCreateInfoEXT, VkDebugReportCallbackEXT, VkInstance

debugReportMessageEXT :: Instance -> DebugReportFlagsEXT -> DebugReportObjectTypeEXT -> Word64 -> CSize -> Int32 -> ByteString -> ByteString -> IO ()

vkDebugReportMessageEXT - Inject a message into a debug stream

Parameters

instance is the debug stream’s VkInstance.

flags specifies the VkDebugReportFlagBitsEXT classification of this event/message.

objectType is a VkDebugReportObjectTypeEXT specifying the type of object being used or created at the time the event was triggered.

object this is the object where the issue was detected. object can be VK_NULL_HANDLE if there is no object associated with the event.

location is an application defined value.

messageCode is an application defined value.

pLayerPrefix is the abbreviation of the component making this event/message.

pMessage is a null-terminated string detailing the trigger conditions.

Description

The call will propagate through the layers and generate callback(s) as indicated by the message’s flags. The parameters are passed on to the callback in addition to the pUserData value that was defined at the time the callback was registered.

Valid Usage

object must be a Vulkan object or VK_NULL_HANDLE

If objectType is not VK_DEBUG_REPORT_OBJECT_TYPE_UNKNOWN_EXT and object is not VK_NULL_HANDLE, object must be a Vulkan object of the corresponding type associated with objectType as defined in https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#debug-report-object-types.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

flags must be a valid combination of VkDebugReportFlagBitsEXT values

flags must not be 0

objectType must be a valid VkDebugReportObjectTypeEXT value

pLayerPrefix must be a null-terminated UTF-8 string

pMessage must be a null-terminated UTF-8 string

See Also

VkDebugReportFlagsEXT, VkDebugReportObjectTypeEXT, VkInstance

destroyDebugReportCallbackEXT :: Instance -> DebugReportCallbackEXT -> Maybe AllocationCallbacks -> IO ()

vkDestroyDebugReportCallbackEXT - Destroy a debug report callback object

Parameters

instance the instance where the callback was created.

callback the VkDebugReportCallbackEXT object to destroy. callback is an externally synchronized object and must not be used on more than one thread at a time. This means that vkDestroyDebugReportCallbackEXT must not be called when a callback is active.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

If VkAllocationCallbacks were provided when callback was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when callback was created, pAllocator must be NULL

Valid Usage (Implicit)

instance must be a valid VkInstance handle

callback must be a valid VkDebugReportCallbackEXT handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

callback must have been created, allocated, or retrieved from instance

Host Synchronization

Host access to callback must be externally synchronized

See Also

VkAllocationCallbacks, VkDebugReportCallbackEXT, VkInstance

withDebugReportCallbackEXT :: Instance -> DebugReportCallbackCreateInfoEXT -> Maybe AllocationCallbacks -> (DebugReportCallbackEXT -> IO a) -> IO a

A safe wrapper for createDebugReportCallbackEXT and destroyDebugReportCallbackEXT using bracket

The allocated value must not be returned from the provided computation

pattern EXT_DEBUG_REPORT_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_DEBUG_REPORT_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_DEBUG_REPORT_CREATE_INFO_EXT :: VkStructureType

pattern STRUCTURE_TYPE_DEBUG_REPORT_CALLBACK_CREATE_INFO_EXT :: a ~ StructureType => a

pattern ERROR_VALIDATION_FAILED_EXT :: a ~ Result => a

pattern OBJECT_TYPE_DEBUG_REPORT_CALLBACK_EXT :: a ~ ObjectType => a