Graphics.Vulkan.Extensions.VK_EXT_debug_utils

Documentation

withCStructDebugUtilsLabelEXT :: DebugUtilsLabelEXT -> (VkDebugUtilsLabelEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDebugUtilsLabelEXT and marshal a DebugUtilsLabelEXT into it. The VkDebugUtilsLabelEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDebugUtilsLabelEXT :: VkDebugUtilsLabelEXT -> IO DebugUtilsLabelEXT

A function to read a VkDebugUtilsLabelEXT and all additional structures in the pointer chain into a DebugUtilsLabelEXT.

data DebugUtilsLabelEXT

VkDebugUtilsLabelEXT - Specify parameters of a label region

Valid Usage (Implicit)

See Also

VkDebugUtilsMessengerCallbackDataEXT, VkStructureType, vkCmdBeginDebugUtilsLabelEXT, vkCmdInsertDebugUtilsLabelEXT, vkQueueBeginDebugUtilsLabelEXT, vkQueueInsertDebugUtilsLabelEXT

Constructors

DebugUtilsLabelEXT

Fields

next :: Maybe SomeVkStruct

labelName :: ByteString

color :: (CFloat, CFloat, CFloat, CFloat)

type DebugUtilsMessageSeverityFlagBitsEXT = VkDebugUtilsMessageSeverityFlagBitsEXT

VkDebugUtilsMessageSeverityFlagBitsEXT - Bitmask specifying which severities of events cause a debug messenger callback

See Also

VkDebugUtilsMessageSeverityFlagsEXT, vkSubmitDebugUtilsMessageEXT

pattern DEBUG_UTILS_MESSAGE_SEVERITY_VERBOSE_BIT_EXT :: a ~ DebugUtilsMessageSeverityFlagBitsEXT => a

VK_DEBUG_UTILS_MESSAGE_SEVERITY_VERBOSE_BIT_EXT specifies the most verbose output indicating all diagnostic messages from the Vulkan loader, layers, and drivers should be captured.

pattern DEBUG_UTILS_MESSAGE_SEVERITY_INFO_BIT_EXT :: a ~ DebugUtilsMessageSeverityFlagBitsEXT => a

VK_DEBUG_UTILS_MESSAGE_SEVERITY_INFO_BIT_EXT specifies an informational message such as resource details that may be handy when debugging an application.

pattern DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT :: a ~ DebugUtilsMessageSeverityFlagBitsEXT => a

VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT specifies use of Vulkan that may expose an app bug. Such cases may not be immediately harmful, such as a fragment shader outputting to a location with no attachment. Other cases may point to behavior that is almost certainly bad when unintended such as using an image whose memory has not been filled. In general if you see a warning but you know that the behavior is intended/desired, then simply ignore the warning.

pattern DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT :: a ~ DebugUtilsMessageSeverityFlagBitsEXT => a

VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT specifies that the application has violated a valid usage condition of the specification.

type DebugUtilsMessageSeverityFlagsEXT = DebugUtilsMessageSeverityFlagBitsEXT

VkDebugUtilsMessageSeverityFlagsEXT - Bitmask of VkDebugUtilsMessageSeverityFlagBitsEXT

Description

VkDebugUtilsMessageSeverityFlagsEXT is a bitmask type for setting a mask of zero or more VkDebugUtilsMessageSeverityFlagBitsEXT.

See Also

VkDebugUtilsMessageSeverityFlagBitsEXT, VkDebugUtilsMessengerCreateInfoEXT

type DebugUtilsMessageTypeFlagBitsEXT = VkDebugUtilsMessageTypeFlagBitsEXT

VkDebugUtilsMessageTypeFlagBitsEXT - Bitmask specifying which types of events cause a debug messenger callback

See Also

VkDebugUtilsMessageTypeFlagsEXT

pattern DEBUG_UTILS_MESSAGE_TYPE_GENERAL_BIT_EXT :: a ~ DebugUtilsMessageTypeFlagBitsEXT => a

VK_DEBUG_UTILS_MESSAGE_TYPE_GENERAL_BIT_EXT specifies that some general event has occurred. This is typically a non-specification, non-performance event.

pattern DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT :: a ~ DebugUtilsMessageTypeFlagBitsEXT => a

VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT specifies that something has occurred during validation against the Vulkan specification that may indicate invalid behavior.

pattern DEBUG_UTILS_MESSAGE_TYPE_PERFORMANCE_BIT_EXT :: a ~ DebugUtilsMessageTypeFlagBitsEXT => a

VK_DEBUG_UTILS_MESSAGE_TYPE_PERFORMANCE_BIT_EXT specifies a potentially non-optimal use of Vulkan, e.g. using vkCmdClearColorImage when setting VkAttachmentDescription::loadOp to VK_ATTACHMENT_LOAD_OP_CLEAR would have worked.

type DebugUtilsMessageTypeFlagsEXT = DebugUtilsMessageTypeFlagBitsEXT

VkDebugUtilsMessageTypeFlagsEXT - Bitmask of VkDebugUtilsMessageTypeFlagBitsEXT

Description

VkDebugUtilsMessageTypeFlagsEXT is a bitmask type for setting a mask of zero or more VkDebugUtilsMessageTypeFlagBitsEXT.

See Also

VkDebugUtilsMessageTypeFlagBitsEXT, VkDebugUtilsMessengerCreateInfoEXT, vkSubmitDebugUtilsMessageEXT

withCStructDebugUtilsMessengerCallbackDataEXT :: DebugUtilsMessengerCallbackDataEXT -> (VkDebugUtilsMessengerCallbackDataEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDebugUtilsMessengerCallbackDataEXT and marshal a DebugUtilsMessengerCallbackDataEXT into it. The VkDebugUtilsMessengerCallbackDataEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDebugUtilsMessengerCallbackDataEXT :: VkDebugUtilsMessengerCallbackDataEXT -> IO DebugUtilsMessengerCallbackDataEXT

A function to read a VkDebugUtilsMessengerCallbackDataEXT and all additional structures in the pointer chain into a DebugUtilsMessengerCallbackDataEXT.

data DebugUtilsMessengerCallbackDataEXT

VkDebugUtilsMessengerCallbackDataEXT - Structure specifying parameters returned to the callback

Description

Note

This structure should only be considered valid during the lifetime of the triggered callback.

Since adding queue and command buffer labels behaves like pushing and popping onto a stack, the order of both pQueueLabels and pCmdBufLabels is based on the order the labels were defined. The result is that the first label in either pQueueLabels or pCmdBufLabels will be the first defined (and therefore the oldest) while the last label in each list will be the most recent.

Note

pQueueLabels will only be non-NULL if one of the objects in pObjects can be related directly to a defined VkQueue which has had one or more labels associated with it.

Likewise, pCmdBufLabels will only be non-NULL if one of the objects in pObjects can be related directly to a defined VkCommandBuffer which has had one or more labels associated with it. Additionally, while command buffer labels allow for beginning and ending across different command buffers, the debug messaging framework cannot guarantee that labels in pCmdBufLables will contain those defined outside of the associated command buffer. This is partially due to the fact that the association of one command buffer with another may not have been defined at the time the debug message is triggered.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DEBUG_UTILS_MESSENGER_CALLBACK_DATA_EXT

pNext must be NULL

flags must be 0

If pMessageIdName is not NULL, pMessageIdName must be a null-terminated UTF-8 string

pMessage must be a null-terminated UTF-8 string

If queueLabelCount is not 0, pQueueLabels must be a valid pointer to an array of queueLabelCount valid VkDebugUtilsLabelEXT structures

If cmdBufLabelCount is not 0, pCmdBufLabels must be a valid pointer to an array of cmdBufLabelCount valid VkDebugUtilsLabelEXT structures

If objectCount is not 0, pObjects must be a valid pointer to an array of objectCount valid VkDebugUtilsObjectNameInfoEXT structures

See Also

VkDebugUtilsLabelEXT, VkDebugUtilsMessengerCallbackDataFlagsEXT, VkDebugUtilsObjectNameInfoEXT, VkStructureType, vkSubmitDebugUtilsMessageEXT

Constructors

DebugUtilsMessengerCallbackDataEXT

Fields

next :: Maybe SomeVkStruct

flags :: DebugUtilsMessengerCallbackDataFlagsEXT

messageIdName :: Maybe ByteString

messageIdNumber :: Int32

message :: ByteString

queueLabels :: Vector DebugUtilsLabelEXT

cmdBufLabels :: Vector DebugUtilsLabelEXT

objects :: Vector DebugUtilsObjectNameInfoEXT

type DebugUtilsMessengerCallbackDataFlagsEXT = VkDebugUtilsMessengerCallbackDataFlagsEXT

type DebugUtilsMessengerCreateFlagsEXT = VkDebugUtilsMessengerCreateFlagsEXT

withCStructDebugUtilsMessengerCreateInfoEXT :: DebugUtilsMessengerCreateInfoEXT -> (VkDebugUtilsMessengerCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDebugUtilsMessengerCreateInfoEXT and marshal a DebugUtilsMessengerCreateInfoEXT into it. The VkDebugUtilsMessengerCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDebugUtilsMessengerCreateInfoEXT :: VkDebugUtilsMessengerCreateInfoEXT -> IO DebugUtilsMessengerCreateInfoEXT

A function to read a VkDebugUtilsMessengerCreateInfoEXT and all additional structures in the pointer chain into a DebugUtilsMessengerCreateInfoEXT.

data DebugUtilsMessengerCreateInfoEXT

VkDebugUtilsMessengerCreateInfoEXT - Structure specifying parameters of a newly created debug messenger

Description

For each VkDebugUtilsMessengerEXT that is created the VkDebugUtilsMessengerCreateInfoEXT::messageSeverity and VkDebugUtilsMessengerCreateInfoEXT::messageType determine when that VkDebugUtilsMessengerCreateInfoEXT::pfnUserCallback is called. The process to determine if the user’s pfnUserCallback is triggered when an event occurs is as follows:

The implementation will perform a bitwise AND of the event’s VkDebugUtilsMessageSeverityFlagBitsEXT with the messageSeverity provided during creation of the VkDebugUtilsMessengerEXT object.

If the value is 0, the message is skipped.

The implementation will perform bitwise AND of the event’s VkDebugUtilsMessageTypeFlagBitsEXT with the messageType provided during the creation of the VkDebugUtilsMessengerEXT object.

If the value is 0, the message is skipped.

The callback will trigger a debug message for the current event

The callback will come directly from the component that detected the event, unless some other layer intercepts the calls for its own purposes (filter them in a different way, log to a system error log, etc.).

An application can receive multiple callbacks if multiple VkDebugUtilsMessengerEXT objects are created. A callback will always be executed in the same thread as the originating Vulkan call.

A callback can be called from multiple threads simultaneously (if the application is making Vulkan calls from multiple threads).

Valid Usage (Implicit)

See Also

PFN_vkDebugUtilsMessengerCallbackEXT, VkDebugUtilsMessageSeverityFlagsEXT, VkDebugUtilsMessageTypeFlagsEXT, VkDebugUtilsMessengerCreateFlagsEXT, VkStructureType, vkCreateDebugUtilsMessengerEXT

Constructors

DebugUtilsMessengerCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

flags :: DebugUtilsMessengerCreateFlagsEXT

messageSeverity :: DebugUtilsMessageSeverityFlagsEXT

messageType :: DebugUtilsMessageTypeFlagsEXT

pfnUserCallback :: PFN_vkDebugUtilsMessengerCallbackEXT

userData :: Ptr ()

type DebugUtilsMessengerEXT = VkDebugUtilsMessengerEXT

VkDebugUtilsMessengerEXT - Opaque handle to a debug messenger object

Description

The debug messenger will provide detailed feedback on the application’s use of Vulkan when events of interest occur. When an event of interest does occur, the debug messenger will submit a debug message to the debug callback that was provided during its creation. Additionally, the debug messenger is responsible with filtering out debug messages that the callback is not interested in and will only provide desired debug messages.

See Also

vkCreateDebugUtilsMessengerEXT, vkDestroyDebugUtilsMessengerEXT

withCStructDebugUtilsObjectNameInfoEXT :: DebugUtilsObjectNameInfoEXT -> (VkDebugUtilsObjectNameInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDebugUtilsObjectNameInfoEXT and marshal a DebugUtilsObjectNameInfoEXT into it. The VkDebugUtilsObjectNameInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDebugUtilsObjectNameInfoEXT :: VkDebugUtilsObjectNameInfoEXT -> IO DebugUtilsObjectNameInfoEXT

A function to read a VkDebugUtilsObjectNameInfoEXT and all additional structures in the pointer chain into a DebugUtilsObjectNameInfoEXT.

data DebugUtilsObjectNameInfoEXT

VkDebugUtilsObjectNameInfoEXT - Specify parameters of a name to give to an object

Description

Applications may change the name associated with an object simply by calling vkSetDebugUtilsObjectNameEXT again with a new string. If pObjectName is an empty string, then any previously set name is removed.

Valid Usage

If objectType is VK_OBJECT_TYPE_UNKNOWN, objectHandle must not be VK_NULL_HANDLE

If objectType is not VK_OBJECT_TYPE_UNKNOWN, objectHandle must be VK_NULL_HANDLE or a valid Vulkan handle of the type associated with objectType as defined in the VkObjectType and Vulkan Handle Relationship table

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DEBUG_UTILS_OBJECT_NAME_INFO_EXT

pNext must be NULL

objectType must be a valid VkObjectType value

If pObjectName is not NULL, pObjectName must be a null-terminated UTF-8 string

See Also

VkDebugUtilsMessengerCallbackDataEXT, VkObjectType, VkStructureType, vkSetDebugUtilsObjectNameEXT

Constructors

DebugUtilsObjectNameInfoEXT

Fields

next :: Maybe SomeVkStruct

objectType :: ObjectType

objectHandle :: Word64

objectName :: Maybe ByteString

withCStructDebugUtilsObjectTagInfoEXT :: DebugUtilsObjectTagInfoEXT -> (VkDebugUtilsObjectTagInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDebugUtilsObjectTagInfoEXT and marshal a DebugUtilsObjectTagInfoEXT into it. The VkDebugUtilsObjectTagInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDebugUtilsObjectTagInfoEXT :: VkDebugUtilsObjectTagInfoEXT -> IO DebugUtilsObjectTagInfoEXT

A function to read a VkDebugUtilsObjectTagInfoEXT and all additional structures in the pointer chain into a DebugUtilsObjectTagInfoEXT.

data DebugUtilsObjectTagInfoEXT

VkDebugUtilsObjectTagInfoEXT - Specify parameters of a tag to attach to an object

Description

The tagName parameter gives a name or identifier to the type of data being tagged. This can be used by debugging layers to easily filter for only data that can be used by that implementation.

Valid Usage (Implicit)

See Also

VkObjectType, VkStructureType, vkSetDebugUtilsObjectTagEXT

Constructors

DebugUtilsObjectTagInfoEXT

Fields

next :: Maybe SomeVkStruct

objectType :: ObjectType

objectHandle :: Word64

tagName :: Word64

tag :: ByteString

cmdBeginDebugUtilsLabelEXT :: CommandBuffer -> DebugUtilsLabelEXT -> IO ()

vkCmdBeginDebugUtilsLabelEXT - Open a command buffer debug label region

Parameters

commandBuffer is the command buffer into which the command is recorded.

pLabelInfo is a pointer to an instance of the VkDebugUtilsLabelEXT structure specifying the parameters of the label region to open.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pLabelInfo must be a valid pointer to a valid VkDebugUtilsLabelEXT structure

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkDebugUtilsLabelEXT

cmdEndDebugUtilsLabelEXT :: CommandBuffer -> IO ()

vkCmdEndDebugUtilsLabelEXT - Close a command buffer label region

Parameters

commandBuffer is the command buffer into which the command is recorded.

Description

An application may open a debug label region in one command buffer and close it in another, or otherwise split debug label regions across multiple command buffers or multiple queue submissions. When viewed from the linear series of submissions to a single queue, the calls to vkCmdBeginDebugUtilsLabelEXT and vkCmdEndDebugUtilsLabelEXT must be matched and balanced.

Valid Usage

There must be an outstanding vkCmdBeginDebugUtilsLabelEXT command prior to the vkCmdEndDebugUtilsLabelEXT on the queue that commandBuffer is submitted to

If commandBuffer is a secondary command buffer, there must be an outstanding vkCmdBeginDebugUtilsLabelEXT command recorded to commandBuffer that has not previously been ended by a call to vkCmdEndDebugUtilsLabelEXT.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer

cmdInsertDebugUtilsLabelEXT :: CommandBuffer -> DebugUtilsLabelEXT -> IO ()

vkCmdInsertDebugUtilsLabelEXT - Insert a label into a command buffer

Parameters

commandBuffer is the command buffer into which the command is recorded.

pInfo is a pointer to an instance of the VkDebugUtilsLabelEXT structure specifying the parameters of the label to insert.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pLabelInfo must be a valid pointer to a valid VkDebugUtilsLabelEXT structure

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkDebugUtilsLabelEXT

createDebugUtilsMessengerEXT :: Instance -> DebugUtilsMessengerCreateInfoEXT -> Maybe AllocationCallbacks -> IO DebugUtilsMessengerEXT

vkCreateDebugUtilsMessengerEXT - Create a debug messenger object

Parameters

instance the instance the messenger will be used with.

pCreateInfo points to a VkDebugUtilsMessengerCreateInfoEXT structure which contains the callback pointer as well as defines the conditions under which this messenger will trigger the callback.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pMessenger is a pointer to record the VkDebugUtilsMessengerEXT object created.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkDebugUtilsMessengerCreateInfoEXT structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pMessenger must be a valid pointer to a VkDebugUtilsMessengerEXT handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

The application must ensure that vkCreateDebugUtilsMessengerEXT is not executed in parallel with any Vulkan command that is also called with instance or child of instance as the dispatchable argument.

See Also

VkAllocationCallbacks, VkDebugUtilsMessengerCreateInfoEXT, VkDebugUtilsMessengerEXT, VkInstance

destroyDebugUtilsMessengerEXT :: Instance -> DebugUtilsMessengerEXT -> Maybe AllocationCallbacks -> IO ()

vkDestroyDebugUtilsMessengerEXT - Destroy a debug messenger object

Parameters

instance the instance where the callback was created.

messenger the VkDebugUtilsMessengerEXT object to destroy. messenger is an externally synchronized object and must not be used on more than one thread at a time. This means that vkDestroyDebugUtilsMessengerEXT must not be called when a callback is active.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

If VkAllocationCallbacks were provided when messenger was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when messenger was created, pAllocator must be NULL

Valid Usage (Implicit)

instance must be a valid VkInstance handle

messenger must be a valid VkDebugUtilsMessengerEXT handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

messenger must have been created, allocated, or retrieved from instance

Host Synchronization

Host access to messenger must be externally synchronized

The application must ensure that vkDestroyDebugUtilsMessengerEXT is not executed in parallel with any Vulkan command that is also called with instance or child of instance as the dispatchable argument.

See Also

VkAllocationCallbacks, VkDebugUtilsMessengerEXT, VkInstance

queueBeginDebugUtilsLabelEXT :: Queue -> DebugUtilsLabelEXT -> IO ()

vkQueueBeginDebugUtilsLabelEXT - Open a queue debug label region

Parameters

queue is the queue in which to start a debug label region.

pLabelInfo is a pointer to an instance of the VkDebugUtilsLabelEXT structure specifying the parameters of the label region to open.

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| -               | -               | Any             | -               |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkDebugUtilsLabelEXT, VkQueue

queueEndDebugUtilsLabelEXT :: Queue -> IO ()

vkQueueEndDebugUtilsLabelEXT - Close a queue debug label region

Parameters

queue is the queue in which a debug label region should be closed.

Description

The calls to vkQueueBeginDebugUtilsLabelEXT and vkQueueEndDebugUtilsLabelEXT must be matched and balanced.

Valid Usage

There must be an outstanding vkQueueBeginDebugUtilsLabelEXT command prior to the vkQueueEndDebugUtilsLabelEXT on the queue

Valid Usage (Implicit)

queue must be a valid VkQueue handle

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| -               | -               | Any             | -               |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkQueue

queueInsertDebugUtilsLabelEXT :: Queue -> DebugUtilsLabelEXT -> IO ()

vkQueueInsertDebugUtilsLabelEXT - Insert a label into a queue

Parameters

queue is the queue into which a debug label will be inserted.

pLabelInfo is a pointer to an instance of the VkDebugUtilsLabelEXT structure specifying the parameters of the label to insert.

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| -               | -               | Any             | -               |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkDebugUtilsLabelEXT, VkQueue

setDebugUtilsObjectNameEXT :: Device -> DebugUtilsObjectNameInfoEXT -> IO ()

vkSetDebugUtilsObjectNameEXT - Give a user-friendly name to an object

Parameters

device is the device that created the object.

pNameInfo is a pointer to an instance of the VkDebugUtilsObjectNameInfoEXT structure specifying the parameters of the name to set on the object.

Valid Usage

pNameInfo->objectType must not be VK_OBJECT_TYPE_UNKNOWN

pNameInfo->objectHandle must not be VK_NULL_HANDLE

Valid Usage (Implicit)

device must be a valid VkDevice handle

pNameInfo must be a valid pointer to a valid VkDebugUtilsObjectNameInfoEXT structure

Host Synchronization

Host access to pNameInfo.objectHandle must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDebugUtilsObjectNameInfoEXT, VkDevice

setDebugUtilsObjectTagEXT :: Device -> DebugUtilsObjectTagInfoEXT -> IO ()

vkSetDebugUtilsObjectTagEXT - Attach arbitrary data to an object

Parameters

device is the device that created the object.

pTagInfo is a pointer to an instance of the VkDebugUtilsObjectTagInfoEXT structure specifying the parameters of the tag to attach to the object.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pTagInfo must be a valid pointer to a valid VkDebugUtilsObjectTagInfoEXT structure

Host Synchronization

Host access to pTagInfo.objectHandle must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDebugUtilsObjectTagInfoEXT, VkDevice

submitDebugUtilsMessageEXT :: Instance -> DebugUtilsMessageSeverityFlagBitsEXT -> DebugUtilsMessageTypeFlagsEXT -> DebugUtilsMessengerCallbackDataEXT -> IO ()

vkSubmitDebugUtilsMessageEXT - Inject a message into a debug stream

Parameters

instance is the debug stream’s VkInstance.

messageSeverity is the VkDebugUtilsMessageSeverityFlagBitsEXT severity of this event/message.

messageTypes is a bitmask of VkDebugUtilsMessageTypeFlagBitsEXT specifying which type of event(s) to identify with this message.

pCallbackData contains all the callback related data in the VkDebugUtilsMessengerCallbackDataEXT structure.

Description

The call will propagate through the layers and generate callback(s) as indicated by the message’s flags. The parameters are passed on to the callback in addition to the pUserData value that was defined at the time the messenger was registered.

Valid Usage (Implicit)

See Also

VkDebugUtilsMessageSeverityFlagBitsEXT, VkDebugUtilsMessageTypeFlagsEXT, VkDebugUtilsMessengerCallbackDataEXT, VkInstance

withDebugUtilsMessengerEXT :: Instance -> DebugUtilsMessengerCreateInfoEXT -> Maybe AllocationCallbacks -> (DebugUtilsMessengerEXT -> IO a) -> IO a

A safe wrapper for createDebugUtilsMessengerEXT and destroyDebugUtilsMessengerEXT using bracket

The allocated value must not be returned from the provided computation

pattern EXT_DEBUG_UTILS_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_DEBUG_UTILS_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_DEBUG_UTILS_OBJECT_NAME_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEBUG_UTILS_OBJECT_TAG_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEBUG_UTILS_LABEL_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEBUG_UTILS_MESSENGER_CALLBACK_DATA_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEBUG_UTILS_MESSENGER_CREATE_INFO_EXT :: a ~ StructureType => a

pattern OBJECT_TYPE_DEBUG_UTILS_MESSENGER_EXT :: a ~ ObjectType => a