Graphics.Vulkan.Extensions.VK_EXT_depth_clip_enable

Documentation

withCStructPhysicalDeviceDepthClipEnableFeaturesEXT :: PhysicalDeviceDepthClipEnableFeaturesEXT -> (VkPhysicalDeviceDepthClipEnableFeaturesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceDepthClipEnableFeaturesEXT and marshal a PhysicalDeviceDepthClipEnableFeaturesEXT into it. The VkPhysicalDeviceDepthClipEnableFeaturesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceDepthClipEnableFeaturesEXT :: VkPhysicalDeviceDepthClipEnableFeaturesEXT -> IO PhysicalDeviceDepthClipEnableFeaturesEXT

A function to read a VkPhysicalDeviceDepthClipEnableFeaturesEXT and all additional structures in the pointer chain into a PhysicalDeviceDepthClipEnableFeaturesEXT.

data PhysicalDeviceDepthClipEnableFeaturesEXT

VkPhysicalDeviceDepthClipEnableFeaturesEXT - Structure indicating support for explicit enable of depth clip

Members

The members of the VkPhysicalDeviceDepthClipEnableFeaturesEXT structure describe the following features:

Description

If the VkPhysicalDeviceDepthClipEnableFeaturesEXT structure is included in the pNext chain of VkPhysicalDeviceFeatures2KHR, it is filled with values indicating whether the feature is supported. VkPhysicalDeviceDepthClipEnableFeaturesEXT can also be used in the pNext chain of VkDeviceCreateInfo to enable this feature.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceDepthClipEnableFeaturesEXT

Fields

next :: Maybe SomeVkStruct

depthClipEnable :: Bool

type PipelineRasterizationDepthClipStateCreateFlagsEXT = VkPipelineRasterizationDepthClipStateCreateFlagsEXT

VkPipelineRasterizationDepthClipStateCreateFlagsEXT - Reserved for future use

Description

VkPipelineRasterizationDepthClipStateCreateFlagsEXT is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkPipelineRasterizationDepthClipStateCreateInfoEXT

withCStructPipelineRasterizationDepthClipStateCreateInfoEXT :: PipelineRasterizationDepthClipStateCreateInfoEXT -> (VkPipelineRasterizationDepthClipStateCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPipelineRasterizationDepthClipStateCreateInfoEXT and marshal a PipelineRasterizationDepthClipStateCreateInfoEXT into it. The VkPipelineRasterizationDepthClipStateCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPipelineRasterizationDepthClipStateCreateInfoEXT :: VkPipelineRasterizationDepthClipStateCreateInfoEXT -> IO PipelineRasterizationDepthClipStateCreateInfoEXT

A function to read a VkPipelineRasterizationDepthClipStateCreateInfoEXT and all additional structures in the pointer chain into a PipelineRasterizationDepthClipStateCreateInfoEXT.

data PipelineRasterizationDepthClipStateCreateInfoEXT

VkPipelineRasterizationDepthClipStateCreateInfoEXT - Structure specifying depth clipping state

Valid Usage (Implicit)

See Also

VkBool32, VkPipelineRasterizationDepthClipStateCreateFlagsEXT, VkStructureType

Constructors

PipelineRasterizationDepthClipStateCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

flags :: PipelineRasterizationDepthClipStateCreateFlagsEXT

depthClipEnable :: Bool

pattern EXT_DEPTH_CLIP_ENABLE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_DEPTH_CLIP_ENABLE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_DEPTH_CLIP_ENABLE_FEATURES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PIPELINE_RASTERIZATION_DEPTH_CLIP_STATE_CREATE_INFO_EXT :: a ~ StructureType => a