Graphics.Vulkan.Extensions.VK_EXT_descriptor_indexing

Documentation

type DescriptorBindingFlagBitsEXT = VkDescriptorBindingFlagBitsEXT

VkDescriptorBindingFlagBitsEXT - Bitmask specifying descriptor set layout binding properties

Description

Note

Note that while VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT_EXT and VK_DESCRIPTOR_BINDING_UPDATE_UNUSED_WHILE_PENDING_BIT_EXT both involve updates to descriptor sets after they are bound, VK_DESCRIPTOR_BINDING_UPDATE_UNUSED_WHILE_PENDING_BIT_EXT is a weaker requirement since it is only about descriptors that are not used, whereas VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT_EXT requires the implementation to observe updates to descriptors that are used.

See Also

VkDescriptorBindingFlagsEXT

pattern DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT_EXT :: a ~ DescriptorBindingFlagBitsEXT => a

VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT_EXT indicates that if descriptors in this binding are updated between when the descriptor set is bound in a command buffer and when that command buffer is submitted to a queue, then the submission will use the most recently set descriptors for this binding and the updates do not invalidate the command buffer. Descriptor bindings created with this flag are also partially exempt from the external synchronization requirement in vkUpdateDescriptorSets. They can be updated concurrently with the set being bound to a command buffer in another thread, but not concurrently with the set being reset or freed.

pattern DESCRIPTOR_BINDING_UPDATE_UNUSED_WHILE_PENDING_BIT_EXT :: a ~ DescriptorBindingFlagBitsEXT => a

VK_DESCRIPTOR_BINDING_UPDATE_UNUSED_WHILE_PENDING_BIT_EXT indicates that descriptors in this binding can be updated after a command buffer has bound this descriptor set, or while a command buffer that uses this descriptor set is pending execution, as long as the descriptors that are updated are not used by those command buffers. If VK_DESCRIPTOR_BINDING_PARTIALLY_BOUND_BIT_EXT is also set, then descriptors can be updated as long as they are not dynamically used by any shader invocations. If VK_DESCRIPTOR_BINDING_PARTIALLY_BOUND_BIT_EXT is not set, then descriptors can be updated as long as they are not statically used by any shader invocations.

pattern DESCRIPTOR_BINDING_PARTIALLY_BOUND_BIT_EXT :: a ~ DescriptorBindingFlagBitsEXT => a

VK_DESCRIPTOR_BINDING_PARTIALLY_BOUND_BIT_EXT indicates that descriptors in this binding that are not /dynamically used/ need not contain valid descriptors at the time the descriptors are consumed. A descriptor is dynamically used if any shader invocation executes an instruction that performs any memory access using the descriptor.

pattern DESCRIPTOR_BINDING_VARIABLE_DESCRIPTOR_COUNT_BIT_EXT :: a ~ DescriptorBindingFlagBitsEXT => a

VK_DESCRIPTOR_BINDING_VARIABLE_DESCRIPTOR_COUNT_BIT_EXT indicates that this descriptor binding has a variable size that will be specified when a descriptor set is allocated using this layout. The value of descriptorCount is treated as an upper bound on the size of the binding. This must only be used for the last binding in the descriptor set layout (i.e. the binding with the largest value of binding). For the purposes of counting against limits such as maxDescriptorSet* and maxPerStageDescriptor*, the full value of descriptorCount is counted .

type DescriptorBindingFlagsEXT = DescriptorBindingFlagBitsEXT

VkDescriptorBindingFlagsEXT - Bitmask of VkDescriptorBindingFlagBitsEXT

Description

VkDescriptorBindingFlagsEXT is a bitmask type for setting a mask of zero or more VkDescriptorBindingFlagBitsEXT.

See Also

VkDescriptorBindingFlagBitsEXT, VkDescriptorSetLayoutBindingFlagsCreateInfoEXT

withCStructDescriptorSetLayoutBindingFlagsCreateInfoEXT :: DescriptorSetLayoutBindingFlagsCreateInfoEXT -> (VkDescriptorSetLayoutBindingFlagsCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDescriptorSetLayoutBindingFlagsCreateInfoEXT and marshal a DescriptorSetLayoutBindingFlagsCreateInfoEXT into it. The VkDescriptorSetLayoutBindingFlagsCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDescriptorSetLayoutBindingFlagsCreateInfoEXT :: VkDescriptorSetLayoutBindingFlagsCreateInfoEXT -> IO DescriptorSetLayoutBindingFlagsCreateInfoEXT

A function to read a VkDescriptorSetLayoutBindingFlagsCreateInfoEXT and all additional structures in the pointer chain into a DescriptorSetLayoutBindingFlagsCreateInfoEXT.

data DescriptorSetLayoutBindingFlagsCreateInfoEXT

VkDescriptorSetLayoutBindingFlagsCreateInfoEXT - Structure specifying creation flags for descriptor set layout bindings

Description

If bindingCount is zero or if this structure is not in the pNext chain, the VkDescriptorBindingFlagsEXT for each descriptor set layout binding is considered to be zero. Otherwise, the descriptor set layout binding at VkDescriptorSetLayoutCreateInfo::pBindings[i] uses the flags in pBindingFlags[i].

Valid Usage

If bindingCount is not zero, bindingCount must equal VkDescriptorSetLayoutCreateInfo::bindingCount

If an element of pBindingFlags includes VK_DESCRIPTOR_BINDING_VARIABLE_DESCRIPTOR_COUNT_BIT_EXT, then all other elements of VkDescriptorSetLayoutCreateInfo::pBindings must have a smaller value of binding

If VkPhysicalDeviceDescriptorIndexingFeaturesEXT::descriptorBindingUniformBufferUpdateAfterBind is not enabled, all bindings with descriptor type VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER must not use VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT_EXT

If VkPhysicalDeviceDescriptorIndexingFeaturesEXT::descriptorBindingSampledImageUpdateAfterBind is not enabled, all bindings with descriptor type VK_DESCRIPTOR_TYPE_SAMPLER, VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, or VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE must not use VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT_EXT

If VkPhysicalDeviceDescriptorIndexingFeaturesEXT::descriptorBindingStorageImageUpdateAfterBind is not enabled, all bindings with descriptor type VK_DESCRIPTOR_TYPE_STORAGE_IMAGE must not use VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT_EXT

If VkPhysicalDeviceDescriptorIndexingFeaturesEXT::descriptorBindingStorageBufferUpdateAfterBind is not enabled, all bindings with descriptor type VK_DESCRIPTOR_TYPE_STORAGE_BUFFER must not use VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT_EXT

If VkPhysicalDeviceDescriptorIndexingFeaturesEXT::descriptorBindingUniformTexelBufferUpdateAfterBind is not enabled, all bindings with descriptor type VK_DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER must not use VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT_EXT

If VkPhysicalDeviceDescriptorIndexingFeaturesEXT::descriptorBindingStorageTexelBufferUpdateAfterBind is not enabled, all bindings with descriptor type VK_DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER must not use VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT_EXT

All bindings with descriptor type VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, or VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC must not use VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT_EXT

If VkPhysicalDeviceDescriptorIndexingFeaturesEXT::descriptorBindingUpdateUnusedWhilePending is not enabled, all elements of pBindingFlags must not include VK_DESCRIPTOR_BINDING_UPDATE_UNUSED_WHILE_PENDING_BIT_EXT

If VkPhysicalDeviceDescriptorIndexingFeaturesEXT::descriptorBindingPartiallyBound is not enabled, all elements of pBindingFlags must not include VK_DESCRIPTOR_BINDING_PARTIALLY_BOUND_BIT_EXT

If VkPhysicalDeviceDescriptorIndexingFeaturesEXT::descriptorBindingVariableDescriptorCount is not enabled, all elements of pBindingFlags must not include VK_DESCRIPTOR_BINDING_VARIABLE_DESCRIPTOR_COUNT_BIT_EXT

If an element of pBindingFlags includes VK_DESCRIPTOR_BINDING_VARIABLE_DESCRIPTOR_COUNT_BIT_EXT, that element’s descriptorType must not be VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC or VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_BINDING_FLAGS_CREATE_INFO_EXT

If bindingCount is not 0, and pBindingFlags is not NULL, pBindingFlags must be a valid pointer to an array of bindingCount valid combinations of VkDescriptorBindingFlagBitsEXT values

See Also

VkDescriptorBindingFlagsEXT, VkStructureType

Constructors

DescriptorSetLayoutBindingFlagsCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

bindingFlags :: Maybe (Vector DescriptorBindingFlagsEXT)

withCStructDescriptorSetVariableDescriptorCountAllocateInfoEXT :: DescriptorSetVariableDescriptorCountAllocateInfoEXT -> (VkDescriptorSetVariableDescriptorCountAllocateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDescriptorSetVariableDescriptorCountAllocateInfoEXT and marshal a DescriptorSetVariableDescriptorCountAllocateInfoEXT into it. The VkDescriptorSetVariableDescriptorCountAllocateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDescriptorSetVariableDescriptorCountAllocateInfoEXT :: VkDescriptorSetVariableDescriptorCountAllocateInfoEXT -> IO DescriptorSetVariableDescriptorCountAllocateInfoEXT

A function to read a VkDescriptorSetVariableDescriptorCountAllocateInfoEXT and all additional structures in the pointer chain into a DescriptorSetVariableDescriptorCountAllocateInfoEXT.

data DescriptorSetVariableDescriptorCountAllocateInfoEXT

VkDescriptorSetVariableDescriptorCountAllocateInfoEXT - Structure specifying additional allocation parameters for descriptor sets

Description

If descriptorSetCount is zero or this structure is not included in the pNext chain, then the variable lengths are considered to be zero. Otherwise, pDescriptorCounts[i] is the number of descriptors in the variable count descriptor binding in the corresponding descriptor set layout. If VkDescriptorSetAllocateInfo::pSetLayouts[i] does not include a variable count descriptor binding, then pDescriptorCounts[i] is ignored.

Valid Usage

If descriptorSetCount is not zero, descriptorSetCount must equal VkDescriptorSetAllocateInfo::descriptorSetCount

If VkDescriptorSetAllocateInfo::pSetLayouts[i] has a variable descriptor count binding, then pDescriptorCounts[i] must be less than or equal to the descriptor count specified for that binding when the descriptor set layout was created.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DESCRIPTOR_SET_VARIABLE_DESCRIPTOR_COUNT_ALLOCATE_INFO_EXT

If descriptorSetCount is not 0, pDescriptorCounts must be a valid pointer to an array of descriptorSetCount uint32_t values

See Also

VkStructureType

Constructors

DescriptorSetVariableDescriptorCountAllocateInfoEXT

Fields

next :: Maybe SomeVkStruct

descriptorCounts :: Vector Word32

withCStructDescriptorSetVariableDescriptorCountLayoutSupportEXT :: DescriptorSetVariableDescriptorCountLayoutSupportEXT -> (VkDescriptorSetVariableDescriptorCountLayoutSupportEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDescriptorSetVariableDescriptorCountLayoutSupportEXT and marshal a DescriptorSetVariableDescriptorCountLayoutSupportEXT into it. The VkDescriptorSetVariableDescriptorCountLayoutSupportEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDescriptorSetVariableDescriptorCountLayoutSupportEXT :: VkDescriptorSetVariableDescriptorCountLayoutSupportEXT -> IO DescriptorSetVariableDescriptorCountLayoutSupportEXT

A function to read a VkDescriptorSetVariableDescriptorCountLayoutSupportEXT and all additional structures in the pointer chain into a DescriptorSetVariableDescriptorCountLayoutSupportEXT.

data DescriptorSetVariableDescriptorCountLayoutSupportEXT

VkDescriptorSetVariableDescriptorCountLayoutSupportEXT - Structure returning information about whether a descriptor set layout can be supported

Description

If the create info includes a variable-sized descriptor, then supported is determined assuming the requested size of the variable-sized descriptor, and maxVariableDescriptorCount is set to the maximum size of that descriptor that can be successfully created (which is greater than or equal to the requested size passed in). If the create info does not include a variable-sized descriptor or if the VkPhysicalDeviceDescriptorIndexingFeaturesEXT::descriptorBindingVariableDescriptorCount feature is not enabled, then maxVariableDescriptorCount is set to zero. For the purposes of this command, a variable-sized descriptor binding with a descriptorCount of zero is treated as if the descriptorCount is one, and thus the binding is not ignored and the maximum descriptor count will be returned. If the layout is not supported, then the value written to maxVariableDescriptorCount is undefined.

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

DescriptorSetVariableDescriptorCountLayoutSupportEXT

Fields

next :: Maybe SomeVkStruct

maxVariableDescriptorCount :: Word32

withCStructPhysicalDeviceDescriptorIndexingFeaturesEXT :: PhysicalDeviceDescriptorIndexingFeaturesEXT -> (VkPhysicalDeviceDescriptorIndexingFeaturesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceDescriptorIndexingFeaturesEXT and marshal a PhysicalDeviceDescriptorIndexingFeaturesEXT into it. The VkPhysicalDeviceDescriptorIndexingFeaturesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceDescriptorIndexingFeaturesEXT :: VkPhysicalDeviceDescriptorIndexingFeaturesEXT -> IO PhysicalDeviceDescriptorIndexingFeaturesEXT

A function to read a VkPhysicalDeviceDescriptorIndexingFeaturesEXT and all additional structures in the pointer chain into a PhysicalDeviceDescriptorIndexingFeaturesEXT.

data PhysicalDeviceDescriptorIndexingFeaturesEXT

VkPhysicalDeviceDescriptorIndexingFeaturesEXT - Structure describing descriptor indexing features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceDescriptorIndexingFeaturesEXT structure describe the following features:

Description

If the VkPhysicalDeviceDescriptorIndexingFeaturesEXT structure is included in the pNext chain of VkPhysicalDeviceFeatures2KHR, it is filled with values indicating whether each feature is supported. VkPhysicalDeviceDescriptorIndexingFeaturesEXT can also be used in the pNext chain of VkDeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceDescriptorIndexingFeaturesEXT

Fields

next :: Maybe SomeVkStruct

shaderInputAttachmentArrayDynamicIndexing :: Bool

shaderUniformTexelBufferArrayDynamicIndexing :: Bool

shaderStorageTexelBufferArrayDynamicIndexing :: Bool

shaderUniformBufferArrayNonUniformIndexing :: Bool

shaderSampledImageArrayNonUniformIndexing :: Bool

shaderStorageBufferArrayNonUniformIndexing :: Bool

shaderStorageImageArrayNonUniformIndexing :: Bool

shaderInputAttachmentArrayNonUniformIndexing :: Bool

shaderUniformTexelBufferArrayNonUniformIndexing :: Bool

shaderStorageTexelBufferArrayNonUniformIndexing :: Bool

descriptorBindingUniformBufferUpdateAfterBind :: Bool

descriptorBindingSampledImageUpdateAfterBind :: Bool

descriptorBindingStorageImageUpdateAfterBind :: Bool

descriptorBindingStorageBufferUpdateAfterBind :: Bool

descriptorBindingUniformTexelBufferUpdateAfterBind :: Bool

descriptorBindingStorageTexelBufferUpdateAfterBind :: Bool

descriptorBindingUpdateUnusedWhilePending :: Bool

descriptorBindingPartiallyBound :: Bool

descriptorBindingVariableDescriptorCount :: Bool

runtimeDescriptorArray :: Bool

withCStructPhysicalDeviceDescriptorIndexingPropertiesEXT :: PhysicalDeviceDescriptorIndexingPropertiesEXT -> (VkPhysicalDeviceDescriptorIndexingPropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceDescriptorIndexingPropertiesEXT and marshal a PhysicalDeviceDescriptorIndexingPropertiesEXT into it. The VkPhysicalDeviceDescriptorIndexingPropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceDescriptorIndexingPropertiesEXT :: VkPhysicalDeviceDescriptorIndexingPropertiesEXT -> IO PhysicalDeviceDescriptorIndexingPropertiesEXT

A function to read a VkPhysicalDeviceDescriptorIndexingPropertiesEXT and all additional structures in the pointer chain into a PhysicalDeviceDescriptorIndexingPropertiesEXT.

data PhysicalDeviceDescriptorIndexingPropertiesEXT

VkPhysicalDeviceDescriptorIndexingPropertiesEXT - Structure describing descriptor indexing properties that can be supported by an implementation

Members

The members of the VkPhysicalDeviceDescriptorIndexingPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceDescriptorIndexingPropertiesEXT structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceDescriptorIndexingPropertiesEXT

Fields

next :: Maybe SomeVkStruct

maxUpdateAfterBindDescriptorsInAllPools :: Word32

shaderUniformBufferArrayNonUniformIndexingNative :: Bool

shaderSampledImageArrayNonUniformIndexingNative :: Bool

shaderStorageBufferArrayNonUniformIndexingNative :: Bool

shaderStorageImageArrayNonUniformIndexingNative :: Bool

shaderInputAttachmentArrayNonUniformIndexingNative :: Bool

robustBufferAccessUpdateAfterBind :: Bool

quadDivergentImplicitLod :: Bool

maxPerStageDescriptorUpdateAfterBindSamplers :: Word32

maxPerStageDescriptorUpdateAfterBindUniformBuffers :: Word32

maxPerStageDescriptorUpdateAfterBindStorageBuffers :: Word32

maxPerStageDescriptorUpdateAfterBindSampledImages :: Word32

maxPerStageDescriptorUpdateAfterBindStorageImages :: Word32

maxPerStageDescriptorUpdateAfterBindInputAttachments :: Word32

maxPerStageUpdateAfterBindResources :: Word32

maxDescriptorSetUpdateAfterBindSamplers :: Word32

maxDescriptorSetUpdateAfterBindUniformBuffers :: Word32

maxDescriptorSetUpdateAfterBindUniformBuffersDynamic :: Word32

maxDescriptorSetUpdateAfterBindStorageBuffers :: Word32

maxDescriptorSetUpdateAfterBindStorageBuffersDynamic :: Word32

maxDescriptorSetUpdateAfterBindSampledImages :: Word32

maxDescriptorSetUpdateAfterBindStorageImages :: Word32

maxDescriptorSetUpdateAfterBindInputAttachments :: Word32

pattern EXT_DESCRIPTOR_INDEXING_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_DESCRIPTOR_INDEXING_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_BINDING_FLAGS_CREATE_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_DESCRIPTOR_INDEXING_FEATURES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_DESCRIPTOR_INDEXING_PROPERTIES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DESCRIPTOR_SET_VARIABLE_DESCRIPTOR_COUNT_ALLOCATE_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DESCRIPTOR_SET_VARIABLE_DESCRIPTOR_COUNT_LAYOUT_SUPPORT_EXT :: a ~ StructureType => a

pattern DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT_EXT :: a ~ DescriptorPoolCreateFlagBits => a

pattern DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT_EXT :: a ~ DescriptorSetLayoutCreateFlagBits => a

pattern ERROR_FRAGMENTATION_EXT :: a ~ Result => a