Graphics.Vulkan.Extensions.VK_EXT_discard_rectangles

Documentation

type DiscardRectangleModeEXT = VkDiscardRectangleModeEXT

VkDiscardRectangleModeEXT - Specify the discard rectangle mode

See Also

VkPipelineDiscardRectangleStateCreateInfoEXT

pattern DISCARD_RECTANGLE_MODE_INCLUSIVE_EXT :: a ~ DiscardRectangleModeEXT => a

VK_DISCARD_RECTANGLE_MODE_INCLUSIVE_EXT specifies that a fragment within any discard rectangle satisfies the test.

pattern DISCARD_RECTANGLE_MODE_EXCLUSIVE_EXT :: a ~ DiscardRectangleModeEXT => a

VK_DISCARD_RECTANGLE_MODE_EXCLUSIVE_EXT specifies that a fragment not within any of the discard rectangles satisfies the test.

withCStructPhysicalDeviceDiscardRectanglePropertiesEXT :: PhysicalDeviceDiscardRectanglePropertiesEXT -> (VkPhysicalDeviceDiscardRectanglePropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceDiscardRectanglePropertiesEXT and marshal a PhysicalDeviceDiscardRectanglePropertiesEXT into it. The VkPhysicalDeviceDiscardRectanglePropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceDiscardRectanglePropertiesEXT :: VkPhysicalDeviceDiscardRectanglePropertiesEXT -> IO PhysicalDeviceDiscardRectanglePropertiesEXT

A function to read a VkPhysicalDeviceDiscardRectanglePropertiesEXT and all additional structures in the pointer chain into a PhysicalDeviceDiscardRectanglePropertiesEXT.

data PhysicalDeviceDiscardRectanglePropertiesEXT

VkPhysicalDeviceDiscardRectanglePropertiesEXT - Structure describing discard rectangle limits that can be supported by an implementation

Members

The members of the VkPhysicalDeviceDiscardRectanglePropertiesEXT structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceDiscardRectanglePropertiesEXT structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

PhysicalDeviceDiscardRectanglePropertiesEXT

Fields

next :: Maybe SomeVkStruct

maxDiscardRectangles :: Word32

type PipelineDiscardRectangleStateCreateFlagsEXT = VkPipelineDiscardRectangleStateCreateFlagsEXT

VkPipelineDiscardRectangleStateCreateFlagsEXT - Reserved for future use

Description

VkPipelineDiscardRectangleStateCreateFlagsEXT is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkPipelineDiscardRectangleStateCreateInfoEXT

withCStructPipelineDiscardRectangleStateCreateInfoEXT :: PipelineDiscardRectangleStateCreateInfoEXT -> (VkPipelineDiscardRectangleStateCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPipelineDiscardRectangleStateCreateInfoEXT and marshal a PipelineDiscardRectangleStateCreateInfoEXT into it. The VkPipelineDiscardRectangleStateCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPipelineDiscardRectangleStateCreateInfoEXT :: VkPipelineDiscardRectangleStateCreateInfoEXT -> IO PipelineDiscardRectangleStateCreateInfoEXT

A function to read a VkPipelineDiscardRectangleStateCreateInfoEXT and all additional structures in the pointer chain into a PipelineDiscardRectangleStateCreateInfoEXT.

data PipelineDiscardRectangleStateCreateInfoEXT

VkPipelineDiscardRectangleStateCreateInfoEXT - Structure specifying discard rectangle

Valid Usage (Implicit)

See Also

VkDiscardRectangleModeEXT, VkPipelineDiscardRectangleStateCreateFlagsEXT, VkRect2D, VkStructureType

Constructors

PipelineDiscardRectangleStateCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

flags :: PipelineDiscardRectangleStateCreateFlagsEXT

discardRectangleMode :: DiscardRectangleModeEXT

discardRectangles :: Maybe (Vector Rect2D)

cmdSetDiscardRectangleEXT :: CommandBuffer -> Word32 -> Vector Rect2D -> IO ()

vkCmdSetDiscardRectangleEXT - Set discard rectangles dynamically

Parameters

commandBuffer is the command buffer into which the command will be recorded.

firstDiscardRectangle is the index of the first discard rectangle whose state is updated by the command.

discardRectangleCount is the number of discard rectangles whose state are updated by the command.

pDiscardRectangles is a pointer to an array of VkRect2D structures specifying discard rectangles.

Description

The discard rectangle taken from element i of pDiscardRectangles replace the current state for the discard rectangle index firstDiscardRectangle + i, for i in [0, discardRectangleCount).

Valid Usage

The bound graphics pipeline must have been created with the VK_DYNAMIC_STATE_DISCARD_RECTANGLE_EXT dynamic state enabled

The sum of firstDiscardRectangle and discardRectangleCount must be less than or equal to VkPhysicalDeviceDiscardRectanglePropertiesEXT::maxDiscardRectangles

The x and y member of offset in each VkRect2D element of pDiscardRectangles must be greater than or equal to 0

Evaluation of (offset.x + extent.width) in each VkRect2D element of pDiscardRectangles must not cause a signed integer addition overflow

Evaluation of (offset.y + extent.height) in each VkRect2D element of pDiscardRectangles must not cause a signed integer addition overflow

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pDiscardRectangles must be a valid pointer to an array of discardRectangleCount VkRect2D structures

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics operations

discardRectangleCount must be greater than 0

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkRect2D

pattern EXT_DISCARD_RECTANGLES_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_DISCARD_RECTANGLES_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_DISCARD_RECTANGLE_PROPERTIES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PIPELINE_DISCARD_RECTANGLE_STATE_CREATE_INFO_EXT :: a ~ StructureType => a

pattern DYNAMIC_STATE_DISCARD_RECTANGLE_EXT :: a ~ DynamicState => a