Graphics.Vulkan.Extensions.VK_EXT_display_control

Documentation

withCStructDeviceEventInfoEXT :: DeviceEventInfoEXT -> (VkDeviceEventInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDeviceEventInfoEXT and marshal a DeviceEventInfoEXT into it. The VkDeviceEventInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDeviceEventInfoEXT :: VkDeviceEventInfoEXT -> IO DeviceEventInfoEXT

A function to read a VkDeviceEventInfoEXT and all additional structures in the pointer chain into a DeviceEventInfoEXT.

data DeviceEventInfoEXT

VkDeviceEventInfoEXT - Describe a device event to create

Valid Usage (Implicit)

See Also

VkDeviceEventTypeEXT, VkStructureType, vkRegisterDeviceEventEXT

Constructors

DeviceEventInfoEXT

Fields

next :: Maybe SomeVkStruct

deviceEvent :: DeviceEventTypeEXT

type DeviceEventTypeEXT = VkDeviceEventTypeEXT

VkDeviceEventTypeEXT - Events that can occur on a device object

See Also

VkDeviceEventInfoEXT

pattern DEVICE_EVENT_TYPE_DISPLAY_HOTPLUG_EXT :: a ~ DeviceEventTypeEXT => a

VK_DEVICE_EVENT_TYPE_DISPLAY_HOTPLUG_EXT specifies that the fence is signaled when a display is plugged into or unplugged from the specified device. Applications can use this notification to determine when they need to re-enumerate the available displays on a device.

withCStructDisplayEventInfoEXT :: DisplayEventInfoEXT -> (VkDisplayEventInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDisplayEventInfoEXT and marshal a DisplayEventInfoEXT into it. The VkDisplayEventInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDisplayEventInfoEXT :: VkDisplayEventInfoEXT -> IO DisplayEventInfoEXT

A function to read a VkDisplayEventInfoEXT and all additional structures in the pointer chain into a DisplayEventInfoEXT.

data DisplayEventInfoEXT

VkDisplayEventInfoEXT - Describe a display event to create

Valid Usage (Implicit)

See Also

VkDisplayEventTypeEXT, VkStructureType, vkRegisterDisplayEventEXT

Constructors

DisplayEventInfoEXT

Fields

next :: Maybe SomeVkStruct

displayEvent :: DisplayEventTypeEXT

type DisplayEventTypeEXT = VkDisplayEventTypeEXT

VkDisplayEventTypeEXT - Events that can occur on a display object

See Also

VkDisplayEventInfoEXT

pattern DISPLAY_EVENT_TYPE_FIRST_PIXEL_OUT_EXT :: a ~ DisplayEventTypeEXT => a

VK_DISPLAY_EVENT_TYPE_FIRST_PIXEL_OUT_EXT specifies that the fence is signaled when the first pixel of the next display refresh cycle leaves the display engine for the display.

withCStructDisplayPowerInfoEXT :: DisplayPowerInfoEXT -> (VkDisplayPowerInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDisplayPowerInfoEXT and marshal a DisplayPowerInfoEXT into it. The VkDisplayPowerInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDisplayPowerInfoEXT :: VkDisplayPowerInfoEXT -> IO DisplayPowerInfoEXT

A function to read a VkDisplayPowerInfoEXT and all additional structures in the pointer chain into a DisplayPowerInfoEXT.

data DisplayPowerInfoEXT

VkDisplayPowerInfoEXT - Describe the power state of a display

Valid Usage (Implicit)

See Also

VkDisplayPowerStateEXT, VkStructureType, vkDisplayPowerControlEXT

Constructors

DisplayPowerInfoEXT

Fields

next :: Maybe SomeVkStruct

powerState :: DisplayPowerStateEXT

type DisplayPowerStateEXT = VkDisplayPowerStateEXT

VkDisplayPowerStateEXT - Possible power states for a display

See Also

VkDisplayPowerInfoEXT

pattern DISPLAY_POWER_STATE_OFF_EXT :: a ~ DisplayPowerStateEXT => a

VK_DISPLAY_POWER_STATE_OFF_EXT specifies that the display is powered down.

pattern DISPLAY_POWER_STATE_SUSPEND_EXT :: a ~ DisplayPowerStateEXT => a

VK_DISPLAY_POWER_STATE_SUSPEND_EXT specifies that the display is put into a low power mode, from which it may be able to transition back to VK_DISPLAY_POWER_STATE_ON_EXT more quickly than if it were in VK_DISPLAY_POWER_STATE_OFF_EXT. This state may be the same as VK_DISPLAY_POWER_STATE_OFF_EXT.

pattern DISPLAY_POWER_STATE_ON_EXT :: a ~ DisplayPowerStateEXT => a

VK_DISPLAY_POWER_STATE_ON_EXT specifies that the display is powered on.

withCStructSwapchainCounterCreateInfoEXT :: SwapchainCounterCreateInfoEXT -> (VkSwapchainCounterCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkSwapchainCounterCreateInfoEXT and marshal a SwapchainCounterCreateInfoEXT into it. The VkSwapchainCounterCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructSwapchainCounterCreateInfoEXT :: VkSwapchainCounterCreateInfoEXT -> IO SwapchainCounterCreateInfoEXT

A function to read a VkSwapchainCounterCreateInfoEXT and all additional structures in the pointer chain into a SwapchainCounterCreateInfoEXT.

data SwapchainCounterCreateInfoEXT

VkSwapchainCounterCreateInfoEXT - Specify the surface counters desired

Valid Usage

The bits in surfaceCounters must be supported by VkSwapchainCreateInfoKHR::surface, as reported by vkGetPhysicalDeviceSurfaceCapabilities2EXT.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_SWAPCHAIN_COUNTER_CREATE_INFO_EXT

surfaceCounters must be a valid combination of VkSurfaceCounterFlagBitsEXT values

See Also

VkStructureType, VkSurfaceCounterFlagsEXT

Constructors

SwapchainCounterCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

surfaceCounters :: SurfaceCounterFlagsEXT

displayPowerControlEXT :: Device -> DisplayKHR -> DisplayPowerInfoEXT -> IO ()

vkDisplayPowerControlEXT - Set the power state of a display

Parameters

device is a logical device associated with display.

display is the display whose power state is modified.

pDisplayPowerInfo is an instance of VkDisplayPowerInfoEXT specifying the new power state of display.

Return Codes

Success

- VK_SUCCESS

See Also

VkDevice, VkDisplayKHR, VkDisplayPowerInfoEXT

getSwapchainCounterEXT :: Device -> SwapchainKHR -> SurfaceCounterFlagBitsEXT -> IO Word64

vkGetSwapchainCounterEXT - Query the current value of a surface counter

Parameters

device is the VkDevice associated with swapchain.

swapchain is the swapchain from which to query the counter value.

counter is the counter to query.

pCounterValue will return the current value of the counter.

Description

If a counter is not available because the swapchain is out of date, the implementation may return VK_ERROR_OUT_OF_DATE_KHR.

Valid Usage

One or more present commands on swapchain must have been processed by the presentation engine.

Valid Usage (Implicit)

device must be a valid VkDevice handle

swapchain must be a valid VkSwapchainKHR handle

counter must be a valid VkSurfaceCounterFlagBitsEXT value

pCounterValue must be a valid pointer to a uint64_t value

Both of device, and swapchain must have been created, allocated, or retrieved from the same VkInstance

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_DEVICE_LOST

VK_ERROR_OUT_OF_DATE_KHR

See Also

VkDevice, VkSurfaceCounterFlagBitsEXT, VkSwapchainKHR

registerDeviceEventEXT :: Device -> DeviceEventInfoEXT -> Maybe AllocationCallbacks -> IO Fence

vkRegisterDeviceEventEXT - Signal a fence when a device event occurs

Parameters

device is a logical device on which the event may occur.

pDeviceEventInfo is a pointer to an instance of the VkDeviceEventInfoEXT structure describing the event of interest to the application.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pFence points to a handle in which the resulting fence object is returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pDeviceEventInfo must be a valid pointer to a valid VkDeviceEventInfoEXT structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pFence must be a valid pointer to a VkFence handle

Return Codes

Success

- VK_SUCCESS

See Also

VkAllocationCallbacks, VkDevice, VkDeviceEventInfoEXT, VkFence

registerDisplayEventEXT :: Device -> DisplayKHR -> DisplayEventInfoEXT -> Maybe AllocationCallbacks -> IO Fence

vkRegisterDisplayEventEXT - Signal a fence when a display event occurs

Parameters

device is a logical device associated with display

display is the display on which the event may occur.

pDisplayEventInfo is a pointer to an instance of the VkDisplayEventInfoEXT structure describing the event of interest to the application.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pFence points to a handle in which the resulting fence object is returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

display must be a valid VkDisplayKHR handle

pDisplayEventInfo must be a valid pointer to a valid VkDisplayEventInfoEXT structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pFence must be a valid pointer to a VkFence handle

Return Codes

Success

- VK_SUCCESS

See Also

VkAllocationCallbacks, VkDevice, VkDisplayEventInfoEXT, VkDisplayKHR, VkFence

pattern EXT_DISPLAY_CONTROL_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_DISPLAY_CONTROL_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_DISPLAY_POWER_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEVICE_EVENT_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DISPLAY_EVENT_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SWAPCHAIN_COUNTER_CREATE_INFO_EXT :: a ~ StructureType => a