Graphics.Vulkan.Extensions.VK_EXT_display_surface_counter

Documentation

withCStructSurfaceCapabilities2EXT :: SurfaceCapabilities2EXT -> (VkSurfaceCapabilities2EXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkSurfaceCapabilities2EXT and marshal a SurfaceCapabilities2EXT into it. The VkSurfaceCapabilities2EXT is only valid inside the provided computation and must not be returned out of it.

fromCStructSurfaceCapabilities2EXT :: VkSurfaceCapabilities2EXT -> IO SurfaceCapabilities2EXT

A function to read a VkSurfaceCapabilities2EXT and all additional structures in the pointer chain into a SurfaceCapabilities2EXT.

data SurfaceCapabilities2EXT

VkSurfaceCapabilities2EXT - Structure describing capabilities of a surface

Members

All members of VkSurfaceCapabilities2EXT are identical to the corresponding members of VkSurfaceCapabilitiesKHR where one exists. The remaining members are:

Valid Usage (Implicit)

See Also

VkCompositeAlphaFlagsKHR, VkExtent2D, VkImageUsageFlags, VkStructureType, VkSurfaceCounterFlagsEXT, VkSurfaceTransformFlagBitsKHR, VkSurfaceTransformFlagsKHR, vkGetPhysicalDeviceSurfaceCapabilities2EXT

Constructors

SurfaceCapabilities2EXT

Fields

next :: Maybe SomeVkStruct

minImageCount :: Word32

maxImageCount :: Word32

currentExtent :: Extent2D

minImageExtent :: Extent2D

maxImageExtent :: Extent2D

maxImageArrayLayers :: Word32

supportedTransforms :: SurfaceTransformFlagsKHR

currentTransform :: SurfaceTransformFlagBitsKHR

supportedCompositeAlpha :: CompositeAlphaFlagsKHR

supportedUsageFlags :: ImageUsageFlags

supportedSurfaceCounters :: SurfaceCounterFlagsEXT

type SurfaceCounterFlagBitsEXT = VkSurfaceCounterFlagBitsEXT

VkSurfaceCounterFlagBitsEXT - Surface-relative counter types

See Also

VkSurfaceCounterFlagsEXT, vkGetSwapchainCounterEXT

pattern SURFACE_COUNTER_VBLANK_EXT :: a ~ SurfaceCounterFlagBitsEXT => a

VK_SURFACE_COUNTER_VBLANK_EXT specifies a counter incrementing once every time a vertical blanking period occurs on the display associated with the surface.

type SurfaceCounterFlagsEXT = SurfaceCounterFlagBitsEXT

VkSurfaceCounterFlagsEXT - Bitmask of VkSurfaceCounterFlagBitsEXT

Description

VkSurfaceCounterFlagsEXT is a bitmask type for setting a mask of zero or more VkSurfaceCounterFlagBitsEXT.

See Also

VkSurfaceCapabilities2EXT, VkSurfaceCounterFlagBitsEXT, VkSwapchainCounterCreateInfoEXT

getPhysicalDeviceSurfaceCapabilities2EXT :: PhysicalDevice -> SurfaceKHR -> IO SurfaceCapabilities2EXT

vkGetPhysicalDeviceSurfaceCapabilities2EXT - Query surface capabilities

Parameters

physicalDevice is the physical device that will be associated with the swapchain to be created, as described for vkCreateSwapchainKHR.

surface is the surface that will be associated with the swapchain.

pSurfaceCapabilities is a pointer to an instance of the VkSurfaceCapabilities2EXT structure in which the capabilities are returned.

Description

vkGetPhysicalDeviceSurfaceCapabilities2EXT behaves similarly to vkGetPhysicalDeviceSurfaceCapabilitiesKHR, with the ability to return extended information by adding extension structures to the pNext chain of its pSurfaceCapabilities parameter.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

surface must be a valid VkSurfaceKHR handle

pSurfaceCapabilities must be a valid pointer to a VkSurfaceCapabilities2EXT structure

Both of physicalDevice, and surface must have been created, allocated, or retrieved from the same VkInstance

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

See Also

VkPhysicalDevice, VkSurfaceCapabilities2EXT, VkSurfaceKHR

pattern EXT_DISPLAY_SURFACE_COUNTER_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_DISPLAY_SURFACE_COUNTER_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_SURFACE_CAPABILITIES2_EXT :: VkStructureType

pattern STRUCTURE_TYPE_SURFACE_CAPABILITIES_2_EXT :: a ~ StructureType => a