Graphics.Vulkan.Extensions.VK_EXT_external_memory_host

Documentation

withCStructImportMemoryHostPointerInfoEXT :: ImportMemoryHostPointerInfoEXT -> (VkImportMemoryHostPointerInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkImportMemoryHostPointerInfoEXT and marshal a ImportMemoryHostPointerInfoEXT into it. The VkImportMemoryHostPointerInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructImportMemoryHostPointerInfoEXT :: VkImportMemoryHostPointerInfoEXT -> IO ImportMemoryHostPointerInfoEXT

A function to read a VkImportMemoryHostPointerInfoEXT and all additional structures in the pointer chain into a ImportMemoryHostPointerInfoEXT.

data ImportMemoryHostPointerInfoEXT

VkImportMemoryHostPointerInfoEXT - import memory from a host pointer

Description

Importing memory from a host pointer shares ownership of the memory between the host and the Vulkan implementation. The application can continue to access the memory through the host pointer but it is the application’s responsibility to synchronize device and non-device access to the underlying memory as defined in Host Access to Device Memory Objects.

Applications can import the same underlying memory into multiple instances of Vulkan and multiple times into a given Vulkan instance. However, implementations may fail to import the same underlying memory multiple times into a given physical device due to platform constraints.

Importing memory from a particular host pointer may not be possible due to additional platform-specific restrictions beyond the scope of this specification in which case the implementation must fail the memory import operation with the error code VK_ERROR_INVALID_EXTERNAL_HANDLE_KHR.

The application must ensure that the imported memory range remains valid and accessible for the lifetime of the imported memory object.

Valid Usage

If handleType is not 0, it must be supported for import, as reported in VkExternalMemoryPropertiesKHR

If handleType is not 0, it must be VK_EXTERNAL_MEMORY_HANDLE_TYPE_HOST_ALLOCATION_BIT_EXT or VK_EXTERNAL_MEMORY_HANDLE_TYPE_HOST_MAPPED_FOREIGN_MEMORY_BIT_EXT

pHostPointer must be a pointer aligned to an integer multiple of VkPhysicalDeviceExternalMemoryHostPropertiesEXT::minImportedHostPointerAlignment

If handleType is VK_EXTERNAL_MEMORY_HANDLE_TYPE_HOST_ALLOCATION_BIT_EXT, pHostPointer must be a pointer to allocationSize number of bytes of host memory, where allocationSize is the member of the VkMemoryAllocateInfo structure this structure is chained to

If handleType is VK_EXTERNAL_MEMORY_HANDLE_TYPE_HOST_MAPPED_FOREIGN_MEMORY_BIT_EXT, pHostPointer must be a pointer to allocationSize number of bytes of host mapped foreign memory, where allocationSize is the member of the VkMemoryAllocateInfo structure this structure is chained to

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMPORT_MEMORY_HOST_POINTER_INFO_EXT

handleType must be a valid VkExternalMemoryHandleTypeFlagBits value

See Also

VkExternalMemoryHandleTypeFlagBits, VkStructureType

Constructors

ImportMemoryHostPointerInfoEXT

Fields

next :: Maybe SomeVkStruct

handleType :: ExternalMemoryHandleTypeFlagBits

hostPointer :: Ptr ()

withCStructMemoryHostPointerPropertiesEXT :: MemoryHostPointerPropertiesEXT -> (VkMemoryHostPointerPropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkMemoryHostPointerPropertiesEXT and marshal a MemoryHostPointerPropertiesEXT into it. The VkMemoryHostPointerPropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructMemoryHostPointerPropertiesEXT :: VkMemoryHostPointerPropertiesEXT -> IO MemoryHostPointerPropertiesEXT

A function to read a VkMemoryHostPointerPropertiesEXT and all additional structures in the pointer chain into a MemoryHostPointerPropertiesEXT.

data MemoryHostPointerPropertiesEXT

VkMemoryHostPointerPropertiesEXT - Properties of external memory host pointer

Description

The value returned by memoryTypeBits must only include bits that identify memory types which are host visible.

Valid Usage (Implicit)

See Also

VkStructureType, vkGetMemoryHostPointerPropertiesEXT

Constructors

MemoryHostPointerPropertiesEXT

Fields

next :: Maybe SomeVkStruct

memoryTypeBits :: Word32

withCStructPhysicalDeviceExternalMemoryHostPropertiesEXT :: PhysicalDeviceExternalMemoryHostPropertiesEXT -> (VkPhysicalDeviceExternalMemoryHostPropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceExternalMemoryHostPropertiesEXT and marshal a PhysicalDeviceExternalMemoryHostPropertiesEXT into it. The VkPhysicalDeviceExternalMemoryHostPropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceExternalMemoryHostPropertiesEXT :: VkPhysicalDeviceExternalMemoryHostPropertiesEXT -> IO PhysicalDeviceExternalMemoryHostPropertiesEXT

A function to read a VkPhysicalDeviceExternalMemoryHostPropertiesEXT and all additional structures in the pointer chain into a PhysicalDeviceExternalMemoryHostPropertiesEXT.

data PhysicalDeviceExternalMemoryHostPropertiesEXT

VkPhysicalDeviceExternalMemoryHostPropertiesEXT - Structure describing external memory host pointer limits that can be supported by an implementation

Members

The members of the VkPhysicalDeviceExternalMemoryHostPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceExternalMemoryHostPropertiesEXT structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkDeviceSize, VkStructureType

Constructors

PhysicalDeviceExternalMemoryHostPropertiesEXT

Fields

next :: Maybe SomeVkStruct

minImportedHostPointerAlignment :: DeviceSize

getMemoryHostPointerPropertiesEXT :: Device -> ExternalMemoryHandleTypeFlagBits -> Ptr () -> IO MemoryHostPointerPropertiesEXT

vkGetMemoryHostPointerPropertiesEXT - Get properties of external memory host pointer

Parameters

device is the logical device that will be importing pHostPointer.

handleType is the type of the handle pHostPointer.

pHostPointer is the host pointer to import from.

pMemoryHostPointerProperties is a pointer to a VkMemoryHostPointerPropertiesEXT structure in which the host pointer properties are returned.

Valid Usage

handleType must be VK_EXTERNAL_MEMORY_HANDLE_TYPE_HOST_ALLOCATION_BIT_EXT or VK_EXTERNAL_MEMORY_HANDLE_TYPE_HOST_MAPPED_FOREIGN_MEMORY_BIT_EXT

pHostPointer must be a pointer aligned to an integer multiple of VkPhysicalDeviceExternalMemoryHostPropertiesEXT::minImportedHostPointerAlignment

If handleType is VK_EXTERNAL_MEMORY_HANDLE_TYPE_HOST_ALLOCATION_BIT_EXT, pHostPointer must be a pointer to host memory

If handleType is VK_EXTERNAL_MEMORY_HANDLE_TYPE_HOST_MAPPED_FOREIGN_MEMORY_BIT_EXT, pHostPointer must be a pointer to host mapped foreign memory

Valid Usage (Implicit)

device must be a valid VkDevice handle

handleType must be a valid VkExternalMemoryHandleTypeFlagBits value

pMemoryHostPointerProperties must be a valid pointer to a VkMemoryHostPointerPropertiesEXT structure

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_INVALID_EXTERNAL_HANDLE

See Also

VkDevice, VkExternalMemoryHandleTypeFlagBits, VkMemoryHostPointerPropertiesEXT

pattern EXT_EXTERNAL_MEMORY_HOST_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_EXTERNAL_MEMORY_HOST_SPEC_VERSION :: Integral a => a

type ExternalMemoryHandleTypeFlagsKHR = ExternalMemoryHandleTypeFlags

type ExternalMemoryHandleTypeFlagBitsKHR = ExternalMemoryHandleTypeFlagBits

pattern STRUCTURE_TYPE_IMPORT_MEMORY_HOST_POINTER_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_MEMORY_HOST_POINTER_PROPERTIES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_MEMORY_HOST_PROPERTIES_EXT :: a ~ StructureType => a

pattern EXTERNAL_MEMORY_HANDLE_TYPE_HOST_ALLOCATION_BIT_EXT :: a ~ ExternalMemoryHandleTypeFlagBits => a

pattern EXTERNAL_MEMORY_HANDLE_TYPE_HOST_MAPPED_FOREIGN_MEMORY_BIT_EXT :: a ~ ExternalMemoryHandleTypeFlagBits => a