Graphics.Vulkan.Extensions.VK_EXT_filter_cubic

Documentation

withCStructFilterCubicImageViewImageFormatPropertiesEXT :: FilterCubicImageViewImageFormatPropertiesEXT -> (VkFilterCubicImageViewImageFormatPropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkFilterCubicImageViewImageFormatPropertiesEXT and marshal a FilterCubicImageViewImageFormatPropertiesEXT into it. The VkFilterCubicImageViewImageFormatPropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructFilterCubicImageViewImageFormatPropertiesEXT :: VkFilterCubicImageViewImageFormatPropertiesEXT -> IO FilterCubicImageViewImageFormatPropertiesEXT

A function to read a VkFilterCubicImageViewImageFormatPropertiesEXT and all additional structures in the pointer chain into a FilterCubicImageViewImageFormatPropertiesEXT.

data FilterCubicImageViewImageFormatPropertiesEXT

VkFilterCubicImageViewImageFormatPropertiesEXT - Structure for querying cubic filtering capabilities of an image view type

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_FILTER_CUBIC_IMAGE_VIEW_IMAGE_FORMAT_PROPERTIES_EXT

Valid Usage

If the pNext chain of the VkImageFormatProperties2 structure contains an instance of VkFilterCubicImageViewImageFormatPropertiesEXT, the pNext chain of the VkPhysicalDeviceImageFormatInfo2 structure must contain an instance of VkPhysicalDeviceImageViewImageFormatInfoEXT with an imageViewType that is compatible with imageType.

See Also

VkBool32, VkStructureType

Constructors

FilterCubicImageViewImageFormatPropertiesEXT

Fields

next :: Maybe SomeVkStruct

filterCubic :: Bool

filterCubicMinmax :: Bool

withCStructPhysicalDeviceImageViewImageFormatInfoEXT :: PhysicalDeviceImageViewImageFormatInfoEXT -> (VkPhysicalDeviceImageViewImageFormatInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceImageViewImageFormatInfoEXT and marshal a PhysicalDeviceImageViewImageFormatInfoEXT into it. The VkPhysicalDeviceImageViewImageFormatInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceImageViewImageFormatInfoEXT :: VkPhysicalDeviceImageViewImageFormatInfoEXT -> IO PhysicalDeviceImageViewImageFormatInfoEXT

A function to read a VkPhysicalDeviceImageViewImageFormatInfoEXT and all additional structures in the pointer chain into a PhysicalDeviceImageViewImageFormatInfoEXT.

data PhysicalDeviceImageViewImageFormatInfoEXT

VkPhysicalDeviceImageViewImageFormatInfoEXT - Structure for providing image view type

Valid Usage (Implicit)

See Also

VkImageViewType, VkStructureType

Constructors

PhysicalDeviceImageViewImageFormatInfoEXT

Fields

next :: Maybe SomeVkStruct

imageViewType :: ImageViewType

pattern EXT_FILTER_CUBIC_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_FILTER_CUBIC_SPEC_VERSION :: Integral a => a

pattern FILTER_CUBIC_EXT :: VkFilter

pattern FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_CUBIC_BIT_EXT :: VkFormatFeatureFlagBits

pattern FILTER_CUBIC_IMG :: a ~ Filter => a

pattern FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_CUBIC_BIT_IMG :: a ~ FormatFeatureFlagBits => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGE_VIEW_IMAGE_FORMAT_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_FILTER_CUBIC_IMAGE_VIEW_IMAGE_FORMAT_PROPERTIES_EXT :: a ~ StructureType => a