Graphics.Vulkan.Extensions.VK_EXT_fragment_density_map

Documentation

withCStructPhysicalDeviceFragmentDensityMapFeaturesEXT :: PhysicalDeviceFragmentDensityMapFeaturesEXT -> (VkPhysicalDeviceFragmentDensityMapFeaturesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceFragmentDensityMapFeaturesEXT and marshal a PhysicalDeviceFragmentDensityMapFeaturesEXT into it. The VkPhysicalDeviceFragmentDensityMapFeaturesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceFragmentDensityMapFeaturesEXT :: VkPhysicalDeviceFragmentDensityMapFeaturesEXT -> IO PhysicalDeviceFragmentDensityMapFeaturesEXT

A function to read a VkPhysicalDeviceFragmentDensityMapFeaturesEXT and all additional structures in the pointer chain into a PhysicalDeviceFragmentDensityMapFeaturesEXT.

data PhysicalDeviceFragmentDensityMapFeaturesEXT

VkPhysicalDeviceFragmentDensityMapFeaturesEXT - Structure describing fragment density map features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceFragmentDensityMapFeaturesEXT structure describe the following features:

Description

If the VkPhysicalDeviceFragmentDensityMapFeaturesEXT structure is included in the pNext chain of VkPhysicalDeviceFeatures2, it is filled with values indicating whether each feature is supported. VkPhysicalDeviceFragmentDensityMapFeaturesEXT can also be used in pNext chain of VkDeviceCreateInfo to enable the features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceFragmentDensityMapFeaturesEXT

Fields

next :: Maybe SomeVkStruct

fragmentDensityMap :: Bool

fragmentDensityMapDynamic :: Bool

fragmentDensityMapNonSubsampledImages :: Bool

withCStructPhysicalDeviceFragmentDensityMapPropertiesEXT :: PhysicalDeviceFragmentDensityMapPropertiesEXT -> (VkPhysicalDeviceFragmentDensityMapPropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceFragmentDensityMapPropertiesEXT and marshal a PhysicalDeviceFragmentDensityMapPropertiesEXT into it. The VkPhysicalDeviceFragmentDensityMapPropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceFragmentDensityMapPropertiesEXT :: VkPhysicalDeviceFragmentDensityMapPropertiesEXT -> IO PhysicalDeviceFragmentDensityMapPropertiesEXT

A function to read a VkPhysicalDeviceFragmentDensityMapPropertiesEXT and all additional structures in the pointer chain into a PhysicalDeviceFragmentDensityMapPropertiesEXT.

data PhysicalDeviceFragmentDensityMapPropertiesEXT

VkPhysicalDeviceFragmentDensityMapPropertiesEXT - Structure describing fragment density map properties that can be supported by an implementation

Members

The members of the VkPhysicalDeviceFragmentDensityMapPropertiesEXT structure describe the following implementation-dependent limits:

Valid Usage (Implicit)

If the VkPhysicalDeviceFragmentDensityMapPropertiesEXT structure is included in the pNext chain of VkPhysicalDeviceProperties2KHR, it is filled with the implementation-dependent limits and properties.

See Also

VkBool32, VkExtent2D, VkStructureType

Constructors

PhysicalDeviceFragmentDensityMapPropertiesEXT

Fields

next :: Maybe SomeVkStruct

minFragmentDensityTexelSize :: Extent2D

maxFragmentDensityTexelSize :: Extent2D

fragmentDensityInvocations :: Bool

withCStructRenderPassFragmentDensityMapCreateInfoEXT :: RenderPassFragmentDensityMapCreateInfoEXT -> (VkRenderPassFragmentDensityMapCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkRenderPassFragmentDensityMapCreateInfoEXT and marshal a RenderPassFragmentDensityMapCreateInfoEXT into it. The VkRenderPassFragmentDensityMapCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructRenderPassFragmentDensityMapCreateInfoEXT :: VkRenderPassFragmentDensityMapCreateInfoEXT -> IO RenderPassFragmentDensityMapCreateInfoEXT

A function to read a VkRenderPassFragmentDensityMapCreateInfoEXT and all additional structures in the pointer chain into a RenderPassFragmentDensityMapCreateInfoEXT.

data RenderPassFragmentDensityMapCreateInfoEXT

VkRenderPassFragmentDensityMapCreateInfoEXT - Structure containing fragment density map attachment for render pass

Description

The fragment density map attachment is read at an implementation-dependent time either by the host during vkCmdBeginRenderPass if the attachment’s image view was not created with flags containing VK_IMAGE_VIEW_CREATE_FRAGMENT_DENSITY_MAP_DYNAMIC_BIT_EXT, or by the device when drawing commands in the renderpass execute VK_PIPELINE_STAGE_FRAGMENT_DENSITY_PROCESS_BIT_EXT.

If this structure is not present, it is as if fragmentDensityMapAttachment was given as VK_ATTACHMENT_UNUSED.

Valid Usage

If fragmentDensityMapAttachment is not VK_ATTACHMENT_UNUSED, fragmentDensityMapAttachment must be less than VkRenderPassCreateInfo::attachmentCount

If fragmentDensityMapAttachment is not VK_ATTACHMENT_UNUSED, fragmentDensityMapAttachment must not be an element of VkSubpassDescription::pInputAttachments, VkSubpassDescription::pColorAttachments, VkSubpassDescription::pResolveAttachments, VkSubpassDescription::pDepthStencilAttachment, or VkSubpassDescription::pPreserveAttachments for any subpass

If fragmentDensityMapAttachment is not VK_ATTACHMENT_UNUSED, layout must be equal to VK_IMAGE_LAYOUT_FRAGMENT_DENSITY_MAP_OPTIMAL_EXT, or VK_IMAGE_LAYOUT_GENERAL

If fragmentDensityMapAttachment is not VK_ATTACHMENT_UNUSED, fragmentDensityMapAttachment must reference an attachment with a loadOp equal to VK_ATTACHMENT_LOAD_OP_LOAD or VK_ATTACHMENT_LOAD_OP_DONT_CARE.

If fragmentDensityMapAttachment is not VK_ATTACHMENT_UNUSED, fragmentDensityMapAttachment must reference an attachment with a storeOp equal to VK_ATTACHMENT_STORE_OP_DONT_CARE.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_RENDER_PASS_FRAGMENT_DENSITY_MAP_CREATE_INFO_EXT

fragmentDensityMapAttachment must be a valid VkAttachmentReference structure

See Also

VkAttachmentReference, VkStructureType

Constructors

RenderPassFragmentDensityMapCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

fragmentDensityMapAttachment :: AttachmentReference

pattern EXT_FRAGMENT_DENSITY_MAP_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_FRAGMENT_DENSITY_MAP_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_FRAGMENT_DENSITY_MAP_FEATURES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_FRAGMENT_DENSITY_MAP_PROPERTIES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_RENDER_PASS_FRAGMENT_DENSITY_MAP_CREATE_INFO_EXT :: a ~ StructureType => a

pattern IMAGE_CREATE_SUBSAMPLED_BIT_EXT :: a ~ ImageCreateFlagBits => a

pattern IMAGE_LAYOUT_FRAGMENT_DENSITY_MAP_OPTIMAL_EXT :: a ~ ImageLayout => a

pattern ACCESS_FRAGMENT_DENSITY_MAP_READ_BIT_EXT :: a ~ AccessFlagBits => a

pattern FORMAT_FEATURE_FRAGMENT_DENSITY_MAP_BIT_EXT :: a ~ FormatFeatureFlagBits => a

pattern IMAGE_USAGE_FRAGMENT_DENSITY_MAP_BIT_EXT :: a ~ ImageUsageFlagBits => a

pattern IMAGE_VIEW_CREATE_FRAGMENT_DENSITY_MAP_DYNAMIC_BIT_EXT :: a ~ ImageViewCreateFlagBits => a

pattern PIPELINE_STAGE_FRAGMENT_DENSITY_PROCESS_BIT_EXT :: a ~ PipelineStageFlagBits => a

pattern SAMPLER_CREATE_SUBSAMPLED_BIT_EXT :: a ~ SamplerCreateFlagBits => a

pattern SAMPLER_CREATE_SUBSAMPLED_COARSE_RECONSTRUCTION_BIT_EXT :: a ~ SamplerCreateFlagBits => a