Graphics.Vulkan.Extensions.VK_EXT_full_screen_exclusive

Documentation

type FullScreenExclusiveEXT = VkFullScreenExclusiveEXT

VkFullScreenExclusiveEXT - Hint values an application can specify affecting full-screen transition behavior

See Also

VkSurfaceFullScreenExclusiveInfoEXT

pattern FULL_SCREEN_EXCLUSIVE_DEFAULT_EXT :: a ~ FullScreenExclusiveEXT => a

VK_FULL_SCREEN_EXCLUSIVE_DEFAULT_EXT indicates the implementation should determine the appropriate full-screen method by whatever means it deems appropriate.

pattern FULL_SCREEN_EXCLUSIVE_ALLOWED_EXT :: a ~ FullScreenExclusiveEXT => a

VK_FULL_SCREEN_EXCLUSIVE_ALLOWED_EXT indicates the implementation may use full-screen exclusive mechanisms when available. Such mechanisms may result in better performance and/or the availability of different presentation capabilities, but may require a more disruptive transition during swapchain initialization, first presentation and/or destruction.

pattern FULL_SCREEN_EXCLUSIVE_DISALLOWED_EXT :: a ~ FullScreenExclusiveEXT => a

VK_FULL_SCREEN_EXCLUSIVE_DISALLOWED_EXT indicates the implementation should avoid using full-screen mechanisms which rely on disruptive transitions.

pattern FULL_SCREEN_EXCLUSIVE_APPLICATION_CONTROLLED_EXT :: a ~ FullScreenExclusiveEXT => a

VK_FULL_SCREEN_EXCLUSIVE_APPLICATION_CONTROLLED_EXT indicates the application will manage full-screen exclusive mode by using the vkAcquireFullScreenExclusiveModeEXT and vkReleaseFullScreenExclusiveModeEXT commands.

withCStructSurfaceCapabilitiesFullScreenExclusiveEXT :: SurfaceCapabilitiesFullScreenExclusiveEXT -> (VkSurfaceCapabilitiesFullScreenExclusiveEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkSurfaceCapabilitiesFullScreenExclusiveEXT and marshal a SurfaceCapabilitiesFullScreenExclusiveEXT into it. The VkSurfaceCapabilitiesFullScreenExclusiveEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructSurfaceCapabilitiesFullScreenExclusiveEXT :: VkSurfaceCapabilitiesFullScreenExclusiveEXT -> IO SurfaceCapabilitiesFullScreenExclusiveEXT

A function to read a VkSurfaceCapabilitiesFullScreenExclusiveEXT and all additional structures in the pointer chain into a SurfaceCapabilitiesFullScreenExclusiveEXT.

data SurfaceCapabilitiesFullScreenExclusiveEXT

VkSurfaceCapabilitiesFullScreenExclusiveEXT - Structure describing full screen exclusive capabilities of a surface

Description

This structure can be included in the pNext chain of VkSurfaceCapabilities2KHR to determine support for exclusive full-screen access. If fullScreenExclusiveSupported is VK_FALSE, it indicates that exclusive full-screen access is not obtainable for this surface.

Applications must not attempt to create swapchains with VK_FULL_SCREEN_EXCLUSIVE_APPLICATION_CONTROLLED_EXT set if fullScreenExclusiveSupported is VK_FALSE.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

SurfaceCapabilitiesFullScreenExclusiveEXT

Fields

next :: Maybe SomeVkStruct

fullScreenExclusiveSupported :: Bool

withCStructSurfaceFullScreenExclusiveInfoEXT :: SurfaceFullScreenExclusiveInfoEXT -> (VkSurfaceFullScreenExclusiveInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkSurfaceFullScreenExclusiveInfoEXT and marshal a SurfaceFullScreenExclusiveInfoEXT into it. The VkSurfaceFullScreenExclusiveInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructSurfaceFullScreenExclusiveInfoEXT :: VkSurfaceFullScreenExclusiveInfoEXT -> IO SurfaceFullScreenExclusiveInfoEXT

A function to read a VkSurfaceFullScreenExclusiveInfoEXT and all additional structures in the pointer chain into a SurfaceFullScreenExclusiveInfoEXT.

data SurfaceFullScreenExclusiveInfoEXT

VkSurfaceFullScreenExclusiveInfoEXT - Structure specifying the preferred full-screen transition behavior

Description

If this structure is not present, fullScreenExclusive is considered to be VK_FULL_SCREEN_EXCLUSIVE_DEFAULT_EXT.

Valid Usage (Implicit)

See Also

VkFullScreenExclusiveEXT, VkStructureType

Constructors

SurfaceFullScreenExclusiveInfoEXT

Fields

next :: Maybe SomeVkStruct

fullScreenExclusive :: FullScreenExclusiveEXT

withCStructSurfaceFullScreenExclusiveWin32InfoEXT :: SurfaceFullScreenExclusiveWin32InfoEXT -> (VkSurfaceFullScreenExclusiveWin32InfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkSurfaceFullScreenExclusiveWin32InfoEXT and marshal a SurfaceFullScreenExclusiveWin32InfoEXT into it. The VkSurfaceFullScreenExclusiveWin32InfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructSurfaceFullScreenExclusiveWin32InfoEXT :: VkSurfaceFullScreenExclusiveWin32InfoEXT -> IO SurfaceFullScreenExclusiveWin32InfoEXT

A function to read a VkSurfaceFullScreenExclusiveWin32InfoEXT and all additional structures in the pointer chain into a SurfaceFullScreenExclusiveWin32InfoEXT.

data SurfaceFullScreenExclusiveWin32InfoEXT

VkSurfaceFullScreenExclusiveWin32InfoEXT - Structure specifying additional creation parameters specific to Win32 fullscreen exclusive mode

Description

Note

If hmonitor is invalidated (e.g. the monitor is unplugged) during the lifetime of a swapchain created with this structure, operations on that swapchain will return VK_ERROR_OUT_OF_DATE_KHR.

Note

It’s the responsibility of the application to change the display settings of the targeted Win32 display using the appropriate platform APIs. Such changes may alter the surface capabilities reported for the created surface.

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

SurfaceFullScreenExclusiveWin32InfoEXT

Fields

next :: Maybe SomeVkStruct

hmonitor :: HMONITOR

acquireFullScreenExclusiveModeEXT :: Device -> SwapchainKHR -> IO ()

vkAcquireFullScreenExclusiveModeEXT - Acquire full-screen exclusive mode for a swapchain

Parameters

device is the device associated with swapchain.

swapchain is the swapchain to acquire exclusive full-screen access for.

Valid Usage

swapchain must not be in the retired state

swapchain must be a swapchain created with an instance of VkSurfaceFullScreenExclusiveInfoEXT, with fullScreenExclusive set to VK_FULL_SCREEN_EXCLUSIVE_APPLICATION_CONTROLLED_EXT

swapchain must not currently have exclusive full-screen access

A return value of VK_SUCCESS indicates that the swapchain successfully acquired exclusive full-screen access. The swapchain will retain this exclusivity until either the application releases exclusive full-screen access with vkReleaseFullScreenExclusiveModeEXT, destroys the swapchain, or if any of the swapchain commands return VK_ERROR_FULL_SCREEN_EXCLUSIVE_MODE_LOST_EXT indicating that the mode was lost because of platform-specific changes.

If the swapchain was unable to acquire exclusive full-screen access to the display then VK_ERROR_INITIALIZATION_FAILED is returned. An application can attempt to acquire exclusive full-screen access again for the same swapchain even if this command fails, or if VK_ERROR_FULL_SCREEN_EXCLUSIVE_MODE_LOST_EXT has been returned by a swapchain command.

Valid Usage (Implicit)

device must be a valid VkDevice handle

swapchain must be a valid VkSwapchainKHR handle

Both of device, and swapchain must have been created, allocated, or retrieved from the same VkInstance

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_INITIALIZATION_FAILED

VK_ERROR_SURFACE_LOST_KHR

See Also

VkDevice, VkSwapchainKHR

getNumPhysicalDeviceSurfacePresentModes2EXT :: PhysicalDevice -> PhysicalDeviceSurfaceInfo2KHR -> IO (VkResult, Word32)

vkGetPhysicalDeviceSurfacePresentModes2EXT - Query supported presentation modes

Parameters

physicalDevice is the physical device that will be associated with the swapchain to be created, as described for vkCreateSwapchainKHR.

pSurfaceInfo points to an instance of the VkPhysicalDeviceSurfaceInfo2KHR structure, describing the surface and other fixed parameters that would be consumed by vkCreateSwapchainKHR.

pPresentModeCount is a pointer to an integer related to the number of presentation modes available or queried, as described below.

pPresentModes is either NULL or a pointer to an array of VkPresentModeKHR values, indicating the supported presentation modes.

Description

vkGetPhysicalDeviceSurfacePresentModes2EXT behaves similarly to vkGetPhysicalDeviceSurfacePresentModesKHR, with the ability to specify extended inputs via chained input structures.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pSurfaceInfo must be a valid pointer to a valid VkPhysicalDeviceSurfaceInfo2KHR structure

pPresentModeCount must be a valid pointer to a uint32_t value

If the value referenced by pPresentModeCount is not 0, and pPresentModes is not NULL, pPresentModes must be a valid pointer to an array of pPresentModeCount VkPresentModeKHR values

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

See Also

VkPhysicalDevice, VkPhysicalDeviceSurfaceInfo2KHR, VkPresentModeKHR

getPhysicalDeviceSurfacePresentModes2EXT :: PhysicalDevice -> PhysicalDeviceSurfaceInfo2KHR -> Word32 -> IO (VkResult, Vector PresentModeKHR)

vkGetPhysicalDeviceSurfacePresentModes2EXT - Query supported presentation modes

Parameters

physicalDevice is the physical device that will be associated with the swapchain to be created, as described for vkCreateSwapchainKHR.

pSurfaceInfo points to an instance of the VkPhysicalDeviceSurfaceInfo2KHR structure, describing the surface and other fixed parameters that would be consumed by vkCreateSwapchainKHR.

pPresentModeCount is a pointer to an integer related to the number of presentation modes available or queried, as described below.

pPresentModes is either NULL or a pointer to an array of VkPresentModeKHR values, indicating the supported presentation modes.

Description

vkGetPhysicalDeviceSurfacePresentModes2EXT behaves similarly to vkGetPhysicalDeviceSurfacePresentModesKHR, with the ability to specify extended inputs via chained input structures.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pSurfaceInfo must be a valid pointer to a valid VkPhysicalDeviceSurfaceInfo2KHR structure

pPresentModeCount must be a valid pointer to a uint32_t value

If the value referenced by pPresentModeCount is not 0, and pPresentModes is not NULL, pPresentModes must be a valid pointer to an array of pPresentModeCount VkPresentModeKHR values

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

See Also

VkPhysicalDevice, VkPhysicalDeviceSurfaceInfo2KHR, VkPresentModeKHR

getAllPhysicalDeviceSurfacePresentModes2EXT :: PhysicalDevice -> PhysicalDeviceSurfaceInfo2KHR -> IO (Vector PresentModeKHR)

Returns all the values available from getPhysicalDeviceSurfacePresentModes2EXT.

releaseFullScreenExclusiveModeEXT :: Device -> SwapchainKHR -> IO ()

vkReleaseFullScreenExclusiveModeEXT - Release full-screen exclusive mode from a swapchain

Parameters

device is the device associated with swapchain.

swapchain is the swapchain to release exclusive full-screen access from.

Description

Note

Applications will not be able to present to swapchain after this call until exclusive full-screen access is reacquired. This is usually useful to handle when an application is minimised or otherwise intends to stop presenting for a time.

Valid Usage

See Also

VkDevice, VkSwapchainKHR

pattern EXT_FULL_SCREEN_EXCLUSIVE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_FULL_SCREEN_EXCLUSIVE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_SURFACE_FULL_SCREEN_EXCLUSIVE_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SURFACE_CAPABILITIES_FULL_SCREEN_EXCLUSIVE_EXT :: a ~ StructureType => a

pattern ERROR_FULL_SCREEN_EXCLUSIVE_MODE_LOST_EXT :: a ~ Result => a

pattern STRUCTURE_TYPE_SURFACE_FULL_SCREEN_EXCLUSIVE_WIN32_INFO_EXT :: a ~ StructureType => a