Graphics.Vulkan.Extensions.VK_EXT_hdr_metadata

Documentation

withCStructHdrMetadataEXT :: HdrMetadataEXT -> (VkHdrMetadataEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkHdrMetadataEXT and marshal a HdrMetadataEXT into it. The VkHdrMetadataEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructHdrMetadataEXT :: VkHdrMetadataEXT -> IO HdrMetadataEXT

A function to read a VkHdrMetadataEXT and all additional structures in the pointer chain into a HdrMetadataEXT.

data HdrMetadataEXT

VkHdrMetadataEXT - structure to specify Hdr metadata

Valid Usage (Implicit)

Note

The validity and use of this data is outside the scope of Vulkan.

See Also

VkStructureType, VkXYColorEXT, vkSetHdrMetadataEXT

Constructors

HdrMetadataEXT

Fields

next :: Maybe SomeVkStruct

displayPrimaryRed :: XYColorEXT

displayPrimaryGreen :: XYColorEXT

displayPrimaryBlue :: XYColorEXT

whitePoint :: XYColorEXT

maxLuminance :: CFloat

minLuminance :: CFloat

maxContentLightLevel :: CFloat

maxFrameAverageLightLevel :: CFloat

withCStructXYColorEXT :: XYColorEXT -> (VkXYColorEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkXYColorEXT and marshal a XYColorEXT into it. The VkXYColorEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructXYColorEXT :: VkXYColorEXT -> IO XYColorEXT

A function to read a VkXYColorEXT and all additional structures in the pointer chain into a XYColorEXT.

data XYColorEXT

VkXYColorEXT - structure to specify X,Y chromaticity coordinates

See Also

VkHdrMetadataEXT

Constructors

XYColorEXT

Fields

x :: CFloat

y :: CFloat

setHdrMetadataEXT :: Device -> Vector SwapchainKHR -> Vector HdrMetadataEXT -> IO ()

vkSetHdrMetadataEXT - function to set Hdr metadata

Parameters

device is the logical device where the swapchain(s) were created.

swapchainCount is the number of swapchains included in pSwapchains.

pSwapchains is a pointer to the array of swapchainCount VkSwapchainKHR handles.

pMetadata is a pointer to the array of swapchainCount VkHdrMetadataEXT structures.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pSwapchains must be a valid pointer to an array of swapchainCount valid VkSwapchainKHR handles

pMetadata must be a valid pointer to an array of swapchainCount valid VkHdrMetadataEXT structures

swapchainCount must be greater than 0

Both of device, and the elements of pSwapchains must have been created, allocated, or retrieved from the same VkInstance

See Also

VkDevice, VkHdrMetadataEXT, VkSwapchainKHR

pattern EXT_HDR_METADATA_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_HDR_METADATA_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_HDR_METADATA_EXT :: a ~ StructureType => a