Graphics.Vulkan.Extensions.VK_EXT_headless_surface

Documentation

type HeadlessSurfaceCreateFlagsEXT = VkHeadlessSurfaceCreateFlagsEXT

withCStructHeadlessSurfaceCreateInfoEXT :: HeadlessSurfaceCreateInfoEXT -> (VkHeadlessSurfaceCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkHeadlessSurfaceCreateInfoEXT and marshal a HeadlessSurfaceCreateInfoEXT into it. The VkHeadlessSurfaceCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructHeadlessSurfaceCreateInfoEXT :: VkHeadlessSurfaceCreateInfoEXT -> IO HeadlessSurfaceCreateInfoEXT

A function to read a VkHeadlessSurfaceCreateInfoEXT and all additional structures in the pointer chain into a HeadlessSurfaceCreateInfoEXT.

data HeadlessSurfaceCreateInfoEXT

VkHeadlessSurfaceCreateInfoEXT - Structure specifying parameters of a newly created headless surface object

Valid Usage (Implicit)

See Also

VkHeadlessSurfaceCreateFlagsEXT, VkStructureType, vkCreateHeadlessSurfaceEXT

Constructors

HeadlessSurfaceCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

flags :: HeadlessSurfaceCreateFlagsEXT

createHeadlessSurfaceEXT :: Instance -> HeadlessSurfaceCreateInfoEXT -> Maybe AllocationCallbacks -> IO SurfaceKHR

vkCreateHeadlessSurfaceEXT - Create a headless VkSurfaceKHR object

Parameters

instance is the instance to associate the surface with.

pCreateInfo is a pointer to an instance of the VkHeadlessSurfaceCreateInfoEXT structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkHeadlessSurfaceCreateInfoEXT structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkHeadlessSurfaceCreateInfoEXT, VkInstance, VkSurfaceKHR

pattern EXT_HEADLESS_SURFACE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_HEADLESS_SURFACE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_HEADLESS_SURFACE_CREATE_INFO_EXT :: a ~ StructureType => a