Graphics.Vulkan.Extensions.VK_EXT_host_query_reset

Documentation

withCStructPhysicalDeviceHostQueryResetFeaturesEXT :: PhysicalDeviceHostQueryResetFeaturesEXT -> (VkPhysicalDeviceHostQueryResetFeaturesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceHostQueryResetFeaturesEXT and marshal a PhysicalDeviceHostQueryResetFeaturesEXT into it. The VkPhysicalDeviceHostQueryResetFeaturesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceHostQueryResetFeaturesEXT :: VkPhysicalDeviceHostQueryResetFeaturesEXT -> IO PhysicalDeviceHostQueryResetFeaturesEXT

A function to read a VkPhysicalDeviceHostQueryResetFeaturesEXT and all additional structures in the pointer chain into a PhysicalDeviceHostQueryResetFeaturesEXT.

data PhysicalDeviceHostQueryResetFeaturesEXT

VkPhysicalDeviceHostQueryResetFeaturesEXT - Structure describing whether queries can be reset from the host

Members

The members of the VkPhysicalDeviceHostQueryResetFeaturesEXT structure describe the following features:

Description

If the VkPhysicalDeviceHostQueryResetFeaturesEXT structure is included in the pNext chain of VkPhysicalDeviceFeatures2, it is filled with values indicating whether the feature is supported. VkPhysicalDeviceHostQueryResetFeaturesEXT can also be used in the pNext chain of VkDeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceHostQueryResetFeaturesEXT

Fields

next :: Maybe SomeVkStruct

hostQueryReset :: Bool

resetQueryPoolEXT :: Device -> QueryPool -> Word32 -> Word32 -> IO ()

vkResetQueryPoolEXT - Reset queries in a query pool

Parameters

queryPool is the handle of the query pool managing the queries being reset.

firstQuery is the initial query index to reset.

queryCount is the number of queries to reset.

Description

This command sets the status of query indices [firstQuery, firstQuery + queryCount - 1] to unavailable.

Valid Usage

The hostQueryReset feature must be enabled

firstQuery must be less than the number of queries in queryPool

The sum of firstQuery and queryCount must be less than or equal to the number of queries in queryPool

Valid Usage (Implicit)

device must be a valid VkDevice handle

queryPool must be a valid VkQueryPool handle

queryPool must have been created, allocated, or retrieved from device

See Also

VkDevice, VkQueryPool

pattern EXT_HOST_QUERY_RESET_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_HOST_QUERY_RESET_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_HOST_QUERY_RESET_FEATURES_EXT :: a ~ StructureType => a