Graphics.Vulkan.Extensions.VK_EXT_image_drm_format_modifier

Documentation

withCStructDrmFormatModifierPropertiesEXT :: DrmFormatModifierPropertiesEXT -> (VkDrmFormatModifierPropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDrmFormatModifierPropertiesEXT and marshal a DrmFormatModifierPropertiesEXT into it. The VkDrmFormatModifierPropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDrmFormatModifierPropertiesEXT :: VkDrmFormatModifierPropertiesEXT -> IO DrmFormatModifierPropertiesEXT

A function to read a VkDrmFormatModifierPropertiesEXT and all additional structures in the pointer chain into a DrmFormatModifierPropertiesEXT.

data DrmFormatModifierPropertiesEXT

VkDrmFormatModifierPropertiesEXT - Structure specifying properties of a format when combined with a DRM format modifier

Description

The returned drmFormatModifierTilingFeatures must contain at least one bit.

The implementation must not return DRM_FORMAT_MOD_INVALID in drmFormatModifier.

An image’s memory planecount (as returned by drmFormatModifierPlaneCount) is distinct from its format planecount (in the sense of multi-planar Y’CBCR formats). In VkImageAspectFlags, each VK_IMAGE_ASPECT_MEMORY_PLANE/i_BIT_EXT represents a _memory plane and each VK_IMAGE_ASPECT_PLANEi_BIT a _format plane/.

An image’s set of format planes is an ordered partition of the image’s content into separable groups of format channels. The ordered partition is encoded in the name of each VkFormat. For example, VK_FORMAT_G8_B8R8_2PLANE_420_UNORM contains two format planes; the first plane contains the green channel and the second plane contains the blue channel and red channel. If the format name does not contain PLANE, then the format contains a single plane; for example, VK_FORMAT_R8G8B8A8_UNORM. Some commands, such as vkCmdCopyBufferToImage, do not operate on all format channels in the image, but instead operate only on the format planes explicitly chosen by the application and operate on each format plane independently.

An image’s set of memory planes is an ordered partition of the image’s memory rather than the image’s content. Each memory plane is a contiguous range of memory. The union of an image’s memory planes is not necessarily contiguous.

If an image is linear, then the partition is the same for memory planes and for /format planes/. Therefore, if the returned drmFormatModifier is DRM_FORMAT_MOD_LINEAR, then drmFormatModifierPlaneCount must equal the format planecount, and drmFormatModifierTilingFeatures must be identical to the VkFormatProperties2::linearTilingFeatures returned in the same pNext chain.

If an image is non-linear, then the partition of the image’s memory into memory planes is implementation-specific and may be unrelated to the partition of the image’s content into format planes. For example, consider an image whose format is VK_FORMAT_G8_B8_R8_3PLANE_420_UNORM, tiling is VK_IMAGE_TILING_DRM_FORMAT_MODIFIER_EXT, whose drmFormatModifier is not DRM_FORMAT_MOD_LINEAR, and flags lacks VK_IMAGE_CREATE_DISJOINT_BIT. The image has 3 format planes, and commands such vkCmdCopyBufferToImage act on each format plane independently as if the data of each /format plane/ were separable from the data of the other planes. In a straightforward implementation, the implementation may store the image’s content in 3 adjacent memory planes where each memory plane corresponds exactly to a format plane. However, the implementation may also store the image’s content in a single memory plane where all format channels are combined using an implementation-private block-compressed format; or the implementation may store the image’s content in a collection of 7 adjacent memory planes using an implementation-private sharding technique. Because the image is non-linear and non-disjoint, the implementation has much freedom when choosing the image’s placement in memory.

The memory planecount applies to function parameters and structures only when the API specifies an explicit requirement on drmFormatModifierPlaneCount. In all other cases, the /memory planecount/ is ignored.

See Also

VkDrmFormatModifierPropertiesListEXT, VkFormatFeatureFlags

Constructors

DrmFormatModifierPropertiesEXT

Fields

drmFormatModifier :: Word64

drmFormatModifierPlaneCount :: Word32

drmFormatModifierTilingFeatures :: FormatFeatureFlags

withCStructDrmFormatModifierPropertiesListEXT :: DrmFormatModifierPropertiesListEXT -> (VkDrmFormatModifierPropertiesListEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDrmFormatModifierPropertiesListEXT and marshal a DrmFormatModifierPropertiesListEXT into it. The VkDrmFormatModifierPropertiesListEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDrmFormatModifierPropertiesListEXT :: VkDrmFormatModifierPropertiesListEXT -> IO DrmFormatModifierPropertiesListEXT

A function to read a VkDrmFormatModifierPropertiesListEXT and all additional structures in the pointer chain into a DrmFormatModifierPropertiesListEXT.

data DrmFormatModifierPropertiesListEXT

VkDrmFormatModifierPropertiesListEXT - Structure specifying the list of DRM format modifiers supported for a format

Description

If pDrmFormatModifierProperties is NULL, then the function returns in drmFormatModifierCount the number of modifiers compatible with the queried format. Otherwise, the application must set drmFormatModifierCount to the length of the array pDrmFormatModifierProperties; the function will write at most drmFormatModifierCount elements to the array, and will return in drmFormatModifierCount the number of elements written.

Among the elements in array pDrmFormatModifierProperties, each returned drmFormatModifier must be unique.

Valid Usage (Implicit)

See Also

VkDrmFormatModifierPropertiesEXT, VkStructureType

Constructors

DrmFormatModifierPropertiesListEXT

Fields

next :: Maybe SomeVkStruct

drmFormatModifierProperties :: Maybe (Vector DrmFormatModifierPropertiesEXT)

withCStructImageDrmFormatModifierExplicitCreateInfoEXT :: ImageDrmFormatModifierExplicitCreateInfoEXT -> (VkImageDrmFormatModifierExplicitCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkImageDrmFormatModifierExplicitCreateInfoEXT and marshal a ImageDrmFormatModifierExplicitCreateInfoEXT into it. The VkImageDrmFormatModifierExplicitCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructImageDrmFormatModifierExplicitCreateInfoEXT :: VkImageDrmFormatModifierExplicitCreateInfoEXT -> IO ImageDrmFormatModifierExplicitCreateInfoEXT

A function to read a VkImageDrmFormatModifierExplicitCreateInfoEXT and all additional structures in the pointer chain into a ImageDrmFormatModifierExplicitCreateInfoEXT.

data ImageDrmFormatModifierExplicitCreateInfoEXT

VkImageDrmFormatModifierExplicitCreateInfoEXT - Specify that an image be created with the provided DRM format modifier and explicit memory layout

Description

The ith member of pPlaneLayouts describes the layout of the image’s ith memory plane (that is, VK_IMAGE_ASPECT_MEMORY_PLANE_i_BIT_EXT). In each element of pPlaneLayouts, the implementation must ignore size. The implementation calculates the size of each plane, which the application can query with vkGetImageSubresourceLayout.

When creating an image with VkImageDrmFormatModifierExplicitCreateInfoEXT, it is the application’s responsibility to satisfy all Valid Usage requirements. However, the implementation must validate that the provided pPlaneLayouts, when combined with the provided drmFormatModifier and other creation parameters in VkImageCreateInfo and its pNext chain, produce a valid image. (This validation is necessarily implementation-dependent and outside the scope of Vulkan, and therefore not described by Valid Usage requirements). If this validation fails, then vkCreateImage returns VK_ERROR_INVALID_DRM_FORMAT_MODIFIER_PLANE_LAYOUT_EXT.

Valid Usage

drmFormatModifier must be compatible with the parameters in VkImageCreateInfo and its pNext chain, as determined by querying VkPhysicalDeviceImageFormatInfo2KHR extended with VkPhysicalDeviceImageDrmFormatModifierInfoEXT.

drmFormatModifierPlaneCount must be equal to the VkDrmFormatModifierPropertiesEXT::drmFormatModifierPlaneCount associated with VkImageCreateInfo::format and drmFormatModifier, as found by querying VkDrmFormatModifierPropertiesListEXT.

For each element of pPlaneLayouts, size must be 0

For each element of pPlaneLayouts, arrayPitch must be 0 if VkImageCreateInfo::arrayLayers is 1.

For each element of pPlaneLayouts, depthPitch must be 0 if VkImageCreateInfo::extent::depth is 1.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMAGE_DRM_FORMAT_MODIFIER_EXPLICIT_CREATE_INFO_EXT

If drmFormatModifierPlaneCount is not 0, pPlaneLayouts must be a valid pointer to an array of drmFormatModifierPlaneCount VkSubresourceLayout structures

See Also

VkStructureType, VkSubresourceLayout

Constructors

ImageDrmFormatModifierExplicitCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

drmFormatModifier :: Word64

planeLayouts :: Vector SubresourceLayout

withCStructImageDrmFormatModifierListCreateInfoEXT :: ImageDrmFormatModifierListCreateInfoEXT -> (VkImageDrmFormatModifierListCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkImageDrmFormatModifierListCreateInfoEXT and marshal a ImageDrmFormatModifierListCreateInfoEXT into it. The VkImageDrmFormatModifierListCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructImageDrmFormatModifierListCreateInfoEXT :: VkImageDrmFormatModifierListCreateInfoEXT -> IO ImageDrmFormatModifierListCreateInfoEXT

A function to read a VkImageDrmFormatModifierListCreateInfoEXT and all additional structures in the pointer chain into a ImageDrmFormatModifierListCreateInfoEXT.

data ImageDrmFormatModifierListCreateInfoEXT

VkImageDrmFormatModifierListCreateInfoEXT - Specify that an image must be created with a DRM format modifier from the provided list

Valid Usage

Each modifier in pDrmFormatModifiers must be compatible with the parameters in VkImageCreateInfo and its pNext chain, as determined by querying VkPhysicalDeviceImageFormatInfo2 extended with VkPhysicalDeviceImageDrmFormatModifierInfoEXT.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMAGE_DRM_FORMAT_MODIFIER_LIST_CREATE_INFO_EXT

pDrmFormatModifiers must be a valid pointer to an array of drmFormatModifierCount uint64_t values

drmFormatModifierCount must be greater than 0

See Also

VkStructureType

Constructors

ImageDrmFormatModifierListCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

drmFormatModifiers :: Vector Word64

withCStructImageDrmFormatModifierPropertiesEXT :: ImageDrmFormatModifierPropertiesEXT -> (VkImageDrmFormatModifierPropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkImageDrmFormatModifierPropertiesEXT and marshal a ImageDrmFormatModifierPropertiesEXT into it. The VkImageDrmFormatModifierPropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructImageDrmFormatModifierPropertiesEXT :: VkImageDrmFormatModifierPropertiesEXT -> IO ImageDrmFormatModifierPropertiesEXT

A function to read a VkImageDrmFormatModifierPropertiesEXT and all additional structures in the pointer chain into a ImageDrmFormatModifierPropertiesEXT.

data ImageDrmFormatModifierPropertiesEXT

VkImageDrmFormatModifierPropertiesEXT - Properties of an image’s Linux DRM format modifier

Description

If the image was created with VkImageDrmFormatModifierListCreateInfoEXT, then the returned drmFormatModifier must belong to the list of modifiers provided at time of image creation in VkImageDrmFormatModifierListCreateInfoEXT::pDrmFormatModifiers. If the image was created with VkImageDrmFormatModifierExplicitCreateInfoEXT, then the returned drmFormatModifier must be the modifier provided at time of image creation in VkImageDrmFormatModifierExplicitCreateInfoEXT::drmFormatModifier.

Valid Usage (Implicit)

See Also

VkStructureType, vkGetImageDrmFormatModifierPropertiesEXT

Constructors

ImageDrmFormatModifierPropertiesEXT

Fields

next :: Maybe SomeVkStruct

drmFormatModifier :: Word64

withCStructPhysicalDeviceImageDrmFormatModifierInfoEXT :: PhysicalDeviceImageDrmFormatModifierInfoEXT -> (VkPhysicalDeviceImageDrmFormatModifierInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceImageDrmFormatModifierInfoEXT and marshal a PhysicalDeviceImageDrmFormatModifierInfoEXT into it. The VkPhysicalDeviceImageDrmFormatModifierInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceImageDrmFormatModifierInfoEXT :: VkPhysicalDeviceImageDrmFormatModifierInfoEXT -> IO PhysicalDeviceImageDrmFormatModifierInfoEXT

A function to read a VkPhysicalDeviceImageDrmFormatModifierInfoEXT and all additional structures in the pointer chain into a PhysicalDeviceImageDrmFormatModifierInfoEXT.

data PhysicalDeviceImageDrmFormatModifierInfoEXT

VkPhysicalDeviceImageDrmFormatModifierInfoEXT - Structure specifying a DRM format modifier as image creation parameter

Description

If the drmFormatModifier is incompatible with the parameters specified in VkPhysicalDeviceImageFormatInfo2 and its pNext chain, then vkGetPhysicalDeviceImageFormatProperties2 returns VK_ERROR_FORMAT_NOT_SUPPORTED. The implementation must support the query of any drmFormatModifier, including unknown and invalid modifier values.

Valid Usage

If sharingMode is VK_SHARING_MODE_CONCURRENT, then pQueueFamilyIndices must be a valid pointer to an array of queueFamilyIndexCount uint32_t values.

If sharingMode is VK_SHARING_MODE_CONCURRENT, then queueFamilyIndexCount must be greater than 1.

If sharingMode is VK_SHARING_MODE_CONCURRENT, each element of pQueueFamilyIndices must be unique and must be less than the pQueueFamilyPropertyCount returned by vkGetPhysicalDeviceQueueFamilyProperties2 for the physicalDevice that was used to create device.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGE_DRM_FORMAT_MODIFIER_INFO_EXT

sharingMode must be a valid VkSharingMode value

See Also

VkSharingMode, VkStructureType

Constructors

PhysicalDeviceImageDrmFormatModifierInfoEXT

Fields

next :: Maybe SomeVkStruct

drmFormatModifier :: Word64

sharingMode :: SharingMode

queueFamilyIndices :: Vector Word32

getImageDrmFormatModifierPropertiesEXT :: Device -> Image -> IO ImageDrmFormatModifierPropertiesEXT

vkGetImageDrmFormatModifierPropertiesEXT - Returns an image’s DRM format modifier

Parameters

device is the logical device that owns the image.

image is the queried image.

pProperties will return properties of the image’s /DRM format modifier/.

Return Codes

Success

- VK_SUCCESS

See Also

VkDevice, VkImage, VkImageDrmFormatModifierPropertiesEXT

pattern EXT_IMAGE_DRM_FORMAT_MODIFIER_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_IMAGE_DRM_FORMAT_MODIFIER_SPEC_VERSION :: Integral a => a

pattern ERROR_INVALID_DRM_FORMAT_MODIFIER_PLANE_LAYOUT_EXT :: a ~ Result => a

pattern STRUCTURE_TYPE_DRM_FORMAT_MODIFIER_PROPERTIES_LIST_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DRM_FORMAT_MODIFIER_PROPERTIES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGE_DRM_FORMAT_MODIFIER_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_IMAGE_DRM_FORMAT_MODIFIER_LIST_CREATE_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_IMAGE_DRM_FORMAT_MODIFIER_EXPLICIT_CREATE_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_IMAGE_DRM_FORMAT_MODIFIER_PROPERTIES_EXT :: a ~ StructureType => a

pattern IMAGE_TILING_DRM_FORMAT_MODIFIER_EXT :: a ~ ImageTiling => a

pattern IMAGE_ASPECT_MEMORY_PLANE_0_BIT_EXT :: a ~ ImageAspectFlagBits => a

pattern IMAGE_ASPECT_MEMORY_PLANE_1_BIT_EXT :: a ~ ImageAspectFlagBits => a

pattern IMAGE_ASPECT_MEMORY_PLANE_2_BIT_EXT :: a ~ ImageAspectFlagBits => a

pattern IMAGE_ASPECT_MEMORY_PLANE_3_BIT_EXT :: a ~ ImageAspectFlagBits => a