Graphics.Vulkan.Extensions.VK_EXT_inline_uniform_block

Documentation

withCStructDescriptorPoolInlineUniformBlockCreateInfoEXT :: DescriptorPoolInlineUniformBlockCreateInfoEXT -> (VkDescriptorPoolInlineUniformBlockCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDescriptorPoolInlineUniformBlockCreateInfoEXT and marshal a DescriptorPoolInlineUniformBlockCreateInfoEXT into it. The VkDescriptorPoolInlineUniformBlockCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDescriptorPoolInlineUniformBlockCreateInfoEXT :: VkDescriptorPoolInlineUniformBlockCreateInfoEXT -> IO DescriptorPoolInlineUniformBlockCreateInfoEXT

A function to read a VkDescriptorPoolInlineUniformBlockCreateInfoEXT and all additional structures in the pointer chain into a DescriptorPoolInlineUniformBlockCreateInfoEXT.

data DescriptorPoolInlineUniformBlockCreateInfoEXT

VkDescriptorPoolInlineUniformBlockCreateInfoEXT - Structure specifying the maximum number of inline uniform block bindings of a newly created descriptor pool

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

DescriptorPoolInlineUniformBlockCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

maxInlineUniformBlockBindings :: Word32

withCStructPhysicalDeviceInlineUniformBlockFeaturesEXT :: PhysicalDeviceInlineUniformBlockFeaturesEXT -> (VkPhysicalDeviceInlineUniformBlockFeaturesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceInlineUniformBlockFeaturesEXT and marshal a PhysicalDeviceInlineUniformBlockFeaturesEXT into it. The VkPhysicalDeviceInlineUniformBlockFeaturesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceInlineUniformBlockFeaturesEXT :: VkPhysicalDeviceInlineUniformBlockFeaturesEXT -> IO PhysicalDeviceInlineUniformBlockFeaturesEXT

A function to read a VkPhysicalDeviceInlineUniformBlockFeaturesEXT and all additional structures in the pointer chain into a PhysicalDeviceInlineUniformBlockFeaturesEXT.

data PhysicalDeviceInlineUniformBlockFeaturesEXT

VkPhysicalDeviceInlineUniformBlockFeaturesEXT - Structure describing inline uniform block features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceInlineUniformBlockFeaturesEXT structure describe the following features:

Description

If the VkPhysicalDeviceInlineUniformBlockFeaturesEXT structure is included in the pNext chain of VkPhysicalDeviceFeatures2, it is filled with values indicating whether each feature is supported. VkPhysicalDeviceInlineUniformBlockFeaturesEXT can also be used in the pNext chain of VkDeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceInlineUniformBlockFeaturesEXT

Fields

next :: Maybe SomeVkStruct

inlineUniformBlock :: Bool

descriptorBindingInlineUniformBlockUpdateAfterBind :: Bool

withCStructPhysicalDeviceInlineUniformBlockPropertiesEXT :: PhysicalDeviceInlineUniformBlockPropertiesEXT -> (VkPhysicalDeviceInlineUniformBlockPropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceInlineUniformBlockPropertiesEXT and marshal a PhysicalDeviceInlineUniformBlockPropertiesEXT into it. The VkPhysicalDeviceInlineUniformBlockPropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceInlineUniformBlockPropertiesEXT :: VkPhysicalDeviceInlineUniformBlockPropertiesEXT -> IO PhysicalDeviceInlineUniformBlockPropertiesEXT

A function to read a VkPhysicalDeviceInlineUniformBlockPropertiesEXT and all additional structures in the pointer chain into a PhysicalDeviceInlineUniformBlockPropertiesEXT.

data PhysicalDeviceInlineUniformBlockPropertiesEXT

VkPhysicalDeviceInlineUniformBlockPropertiesEXT - Structure describing inline uniform block properties that can be supported by an implementation

Members

The members of the VkPhysicalDeviceInlineUniformBlockPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceInlineUniformBlockPropertiesEXT structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

PhysicalDeviceInlineUniformBlockPropertiesEXT

Fields

next :: Maybe SomeVkStruct

maxInlineUniformBlockSize :: Word32

maxPerStageDescriptorInlineUniformBlocks :: Word32

maxPerStageDescriptorUpdateAfterBindInlineUniformBlocks :: Word32

maxDescriptorSetInlineUniformBlocks :: Word32

maxDescriptorSetUpdateAfterBindInlineUniformBlocks :: Word32

withCStructWriteDescriptorSetInlineUniformBlockEXT :: WriteDescriptorSetInlineUniformBlockEXT -> (VkWriteDescriptorSetInlineUniformBlockEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkWriteDescriptorSetInlineUniformBlockEXT and marshal a WriteDescriptorSetInlineUniformBlockEXT into it. The VkWriteDescriptorSetInlineUniformBlockEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructWriteDescriptorSetInlineUniformBlockEXT :: VkWriteDescriptorSetInlineUniformBlockEXT -> IO WriteDescriptorSetInlineUniformBlockEXT

A function to read a VkWriteDescriptorSetInlineUniformBlockEXT and all additional structures in the pointer chain into a WriteDescriptorSetInlineUniformBlockEXT.

data WriteDescriptorSetInlineUniformBlockEXT

VkWriteDescriptorSetInlineUniformBlockEXT - Structure specifying inline uniform block data

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

WriteDescriptorSetInlineUniformBlockEXT

Fields

next :: Maybe SomeVkStruct

data' :: ByteString

pattern EXT_INLINE_UNIFORM_BLOCK_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_INLINE_UNIFORM_BLOCK_SPEC_VERSION :: Integral a => a

pattern DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT :: a ~ DescriptorType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_INLINE_UNIFORM_BLOCK_FEATURES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_INLINE_UNIFORM_BLOCK_PROPERTIES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET_INLINE_UNIFORM_BLOCK_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DESCRIPTOR_POOL_INLINE_UNIFORM_BLOCK_CREATE_INFO_EXT :: a ~ StructureType => a