Graphics.Vulkan.Extensions.VK_EXT_line_rasterization

Documentation

cmdSetLineStippleEXT :: CommandBuffer -> ("lineStippleFactor" ::: Word32) -> ("lineStipplePattern" ::: Word16) -> IO ()

vkCmdSetLineStippleEXT - Set the dynamic line width state

Parameters

CommandBuffer is the command buffer into which the command will be recorded.

lineStippleFactor is the repeat factor used in stippled line rasterization.

lineStipplePattern is the bit pattern used in stippled line rasterization.

Valid Usage

lineStippleFactor must be in the range [1,256]

Valid Usage (Implicit)

CommandBuffer must be a valid CommandBuffer handle

CommandBuffer must be in the recording state

The CommandPool that CommandBuffer was allocated from must support graphics operations

Host Synchronization

Host access to CommandBuffer must be externally synchronized

Host access to the CommandPool that CommandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

CommandBuffer

data PhysicalDeviceLineRasterizationFeaturesEXT

VkPhysicalDeviceLineRasterizationFeaturesEXT - Structure describing the line rasterization features that can be supported by an implementation

Members

The members of the PhysicalDeviceLineRasterizationFeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceLineRasterizationFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether the feature is supported. PhysicalDeviceLineRasterizationFeaturesEXT can also be included in the pNext chain of DeviceCreateInfo to enable the feature.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceLineRasterizationFeaturesEXT

Fields

rectangularLines :: Bool

rectangularLines indicates whether the implementation supports rectangular line rasterization.

bresenhamLines :: Bool

bresenhamLines indicates whether the implementation supports Bresenham-style line rasterization.

smoothLines :: Bool

smoothLines indicates whether the implementation supports smooth line rasterization.

stippledRectangularLines :: Bool

stippledRectangularLines indicates whether the implementation supports stippled line rasterization with LINE_RASTERIZATION_MODE_RECTANGULAR_EXT lines, or with LINE_RASTERIZATION_MODE_DEFAULT_EXT lines if PhysicalDeviceLimits::strictLines is TRUE.

stippledBresenhamLines :: Bool

stippledBresenhamLines indicates whether the implementation supports stippled line rasterization with LINE_RASTERIZATION_MODE_BRESENHAM_EXT lines.

stippledSmoothLines :: Bool

stippledSmoothLines indicates whether the implementation supports stippled line rasterization with LINE_RASTERIZATION_MODE_RECTANGULAR_SMOOTH_EXT lines.

data PhysicalDeviceLineRasterizationPropertiesEXT

VkPhysicalDeviceLineRasterizationPropertiesEXT - Structure describing line rasterization properties supported by an implementation

Members

The members of the PhysicalDeviceLineRasterizationPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the PhysicalDeviceLineRasterizationPropertiesEXT structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

StructureType

Constructors

PhysicalDeviceLineRasterizationPropertiesEXT

Fields

lineSubPixelPrecisionBits :: Word32

lineSubPixelPrecisionBits is the number of bits of subpixel precision in framebuffer coordinates xf and yf when rasterizing line segments.

data PipelineRasterizationLineStateCreateInfoEXT

VkPipelineRasterizationLineStateCreateInfoEXT - Structure specifying parameters of a newly created pipeline line rasterization state

Valid Usage

If lineRasterizationMode is LINE_RASTERIZATION_MODE_RECTANGULAR_EXT, then the rectangularLines feature must be enabled

If lineRasterizationMode is LINE_RASTERIZATION_MODE_BRESENHAM_EXT, then the bresenhamLines feature must be enabled

If lineRasterizationMode is LINE_RASTERIZATION_MODE_RECTANGULAR_SMOOTH_EXT, then the smoothLines feature must be enabled

If stippledLineEnable is TRUE and lineRasterizationMode is LINE_RASTERIZATION_MODE_RECTANGULAR_EXT, then the stippledRectangularLines feature must be enabled

If stippledLineEnable is TRUE and lineRasterizationMode is LINE_RASTERIZATION_MODE_BRESENHAM_EXT, then the stippledBresenhamLines feature must be enabled

If stippledLineEnable is TRUE and lineRasterizationMode is LINE_RASTERIZATION_MODE_RECTANGULAR_SMOOTH_EXT, then the stippledSmoothLines feature must be enabled

If stippledLineEnable is TRUE and lineRasterizationMode is LINE_RASTERIZATION_MODE_DEFAULT_EXT, then the stippledRectangularLines feature must be enabled and PhysicalDeviceLimits::strictLines must be TRUE

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PIPELINE_RASTERIZATION_LINE_STATE_CREATE_INFO_EXT

lineRasterizationMode must be a valid LineRasterizationModeEXT value

See Also

Bool32, LineRasterizationModeEXT, StructureType

Constructors

PipelineRasterizationLineStateCreateInfoEXT

Fields

lineRasterizationMode :: LineRasterizationModeEXT

lineRasterizationMode is a LineRasterizationModeEXT value selecting the style of line rasterization.

stippledLineEnable :: Bool

stippledLineEnable enables stippled line rasterization.

lineStippleFactor :: Word32

lineStippleFactor is the repeat factor used in stippled line rasterization.

lineStipplePattern :: Word16

lineStipplePattern is the bit pattern used in stippled line rasterization.

newtype LineRasterizationModeEXT

VkLineRasterizationModeEXT - Line rasterization modes

See Also

PipelineRasterizationLineStateCreateInfoEXT

Constructors

LineRasterizationModeEXT Int32

Bundled Patterns

pattern LINE_RASTERIZATION_MODE_DEFAULT_EXT :: LineRasterizationModeEXT
LINE_RASTERIZATION_MODE_DEFAULT_EXT is equivalent to LINE_RASTERIZATION_MODE_RECTANGULAR_EXT if PhysicalDeviceLimits::strictLines is TRUE, otherwise lines are drawn as non-strictLines parallelograms. Both of these modes are defined in Basic Line Segment Rasterization.

pattern LINE_RASTERIZATION_MODE_RECTANGULAR_EXT :: LineRasterizationModeEXT
LINE_RASTERIZATION_MODE_RECTANGULAR_EXT specifies lines drawn as if they were rectangles extruded from the line

pattern LINE_RASTERIZATION_MODE_BRESENHAM_EXT :: LineRasterizationModeEXT
LINE_RASTERIZATION_MODE_BRESENHAM_EXT specifies lines drawn by determining which pixel diamonds the line intersects and exits, as defined in Bresenham Line Segment Rasterization.

pattern LINE_RASTERIZATION_MODE_RECTANGULAR_SMOOTH_EXT :: LineRasterizationModeEXT
LINE_RASTERIZATION_MODE_RECTANGULAR_SMOOTH_EXT specifies lines drawn if they were rectangles extruded from the line, with alpha falloff, as defined in Smooth Lines.

type EXT_LINE_RASTERIZATION_SPEC_VERSION = 1

pattern EXT_LINE_RASTERIZATION_SPEC_VERSION :: forall a. Integral a => a

type EXT_LINE_RASTERIZATION_EXTENSION_NAME = "VK_EXT_line_rasterization"

pattern EXT_LINE_RASTERIZATION_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a