Graphics.Vulkan.Extensions.VK_EXT_memory_budget

Documentation

withCStructPhysicalDeviceMemoryBudgetPropertiesEXT :: PhysicalDeviceMemoryBudgetPropertiesEXT -> (VkPhysicalDeviceMemoryBudgetPropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceMemoryBudgetPropertiesEXT and marshal a PhysicalDeviceMemoryBudgetPropertiesEXT into it. The VkPhysicalDeviceMemoryBudgetPropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceMemoryBudgetPropertiesEXT :: VkPhysicalDeviceMemoryBudgetPropertiesEXT -> IO PhysicalDeviceMemoryBudgetPropertiesEXT

A function to read a VkPhysicalDeviceMemoryBudgetPropertiesEXT and all additional structures in the pointer chain into a PhysicalDeviceMemoryBudgetPropertiesEXT.

data PhysicalDeviceMemoryBudgetPropertiesEXT

VkPhysicalDeviceMemoryBudgetPropertiesEXT - Structure specifying physical device memory budget and usage

Description

The values returned in this structure are not invariant. The heapBudget and heapUsage values must be zero for array elements greater than or equal to VkPhysicalDeviceMemoryProperties::memoryHeapCount. The heapBudget value must be non-zero for array elements less than VkPhysicalDeviceMemoryProperties::memoryHeapCount. The heapBudget value must be less than or equal to VkMemoryHeap::size for each heap.

Valid Usage (Implicit)

See Also

VkDeviceSize, VkStructureType

Constructors

PhysicalDeviceMemoryBudgetPropertiesEXT

Fields

next :: Maybe SomeVkStruct

heapBudget :: Vector DeviceSize

heapUsage :: Vector DeviceSize

pattern EXT_MEMORY_BUDGET_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_MEMORY_BUDGET_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_MEMORY_BUDGET_PROPERTIES_EXT :: a ~ StructureType => a