Graphics.Vulkan.Extensions.VK_EXT_memory_priority

Documentation

withCStructMemoryPriorityAllocateInfoEXT :: MemoryPriorityAllocateInfoEXT -> (VkMemoryPriorityAllocateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkMemoryPriorityAllocateInfoEXT and marshal a MemoryPriorityAllocateInfoEXT into it. The VkMemoryPriorityAllocateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructMemoryPriorityAllocateInfoEXT :: VkMemoryPriorityAllocateInfoEXT -> IO MemoryPriorityAllocateInfoEXT

A function to read a VkMemoryPriorityAllocateInfoEXT and all additional structures in the pointer chain into a MemoryPriorityAllocateInfoEXT.

data MemoryPriorityAllocateInfoEXT

VkMemoryPriorityAllocateInfoEXT - Specify a memory allocation priority

Description

Memory allocations with higher priority may be more likely to stay in device-local memory when the system is under memory pressure.

If this structure is not included, it is as if the priority value were 0.5.

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

MemoryPriorityAllocateInfoEXT

Fields

next :: Maybe SomeVkStruct

priority :: CFloat

withCStructPhysicalDeviceMemoryPriorityFeaturesEXT :: PhysicalDeviceMemoryPriorityFeaturesEXT -> (VkPhysicalDeviceMemoryPriorityFeaturesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceMemoryPriorityFeaturesEXT and marshal a PhysicalDeviceMemoryPriorityFeaturesEXT into it. The VkPhysicalDeviceMemoryPriorityFeaturesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceMemoryPriorityFeaturesEXT :: VkPhysicalDeviceMemoryPriorityFeaturesEXT -> IO PhysicalDeviceMemoryPriorityFeaturesEXT

A function to read a VkPhysicalDeviceMemoryPriorityFeaturesEXT and all additional structures in the pointer chain into a PhysicalDeviceMemoryPriorityFeaturesEXT.

data PhysicalDeviceMemoryPriorityFeaturesEXT

VkPhysicalDeviceMemoryPriorityFeaturesEXT - Structure describing memory priority features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceMemoryPriorityFeaturesEXT structure describe the following features:

Description

If the VkPhysicalDeviceMemoryPriorityFeaturesEXT structure is included in the pNext chain of VkPhysicalDeviceFeatures2KHR, it is filled with values indicating whether the feature is supported. VkPhysicalDeviceMemoryPriorityFeaturesEXT can also be used in the pNext chain of VkDeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceMemoryPriorityFeaturesEXT

Fields

next :: Maybe SomeVkStruct

memoryPriority :: Bool

pattern EXT_MEMORY_PRIORITY_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_MEMORY_PRIORITY_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_MEMORY_PRIORITY_FEATURES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_MEMORY_PRIORITY_ALLOCATE_INFO_EXT :: a ~ StructureType => a