Graphics.Vulkan.Extensions.VK_EXT_metal_surface

Documentation

createMetalSurfaceEXT :: Instance -> MetalSurfaceCreateInfoEXT -> ("allocator" ::: Maybe AllocationCallbacks) -> IO SurfaceKHR

vkCreateMetalSurfaceEXT - Create a VkSurfaceKHR object for CAMetalLayer

Parameters

Instance is the instance with which to associate the surface.

pCreateInfo is a pointer to a MetalSurfaceCreateInfoEXT structure specifying parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface is a pointer to a SurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

Instance must be a valid Instance handle

pCreateInfo must be a valid pointer to a valid MetalSurfaceCreateInfoEXT structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pSurface must be a valid pointer to a SurfaceKHR handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_NATIVE_WINDOW_IN_USE_KHR

See Also

AllocationCallbacks, Instance, MetalSurfaceCreateInfoEXT, SurfaceKHR

data MetalSurfaceCreateInfoEXT

VkMetalSurfaceCreateInfoEXT - Structure specifying parameters of a newly created Metal surface object

Valid Usage (Implicit)

See Also

MetalSurfaceCreateFlagsEXT, StructureType, createMetalSurfaceEXT

Constructors

MetalSurfaceCreateInfoEXT

Fields

flags :: MetalSurfaceCreateFlagsEXT

Flags must be 0

layer :: Ptr CAMetalLayer

pLayer is a reference to a CAMetalLayer object representing a renderable surface.

newtype MetalSurfaceCreateFlagsEXT

Constructors

MetalSurfaceCreateFlagsEXT Flags

type EXT_METAL_SURFACE_SPEC_VERSION = 1

pattern EXT_METAL_SURFACE_SPEC_VERSION :: forall a. Integral a => a

type EXT_METAL_SURFACE_EXTENSION_NAME = "VK_EXT_metal_surface"

pattern EXT_METAL_SURFACE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype SurfaceKHR

VkSurfaceKHR - Opaque handle to a surface object

Description

The VK_KHR_surface extension declares the SurfaceKHR object, and provides a function for destroying SurfaceKHR objects. Separate platform-specific extensions each provide a function for creating a SurfaceKHR object for the respective platform. From the application’s perspective this is an opaque handle, just like the handles of other Vulkan objects.

See Also

PhysicalDeviceSurfaceInfo2KHR, SwapchainCreateInfoKHR, createAndroidSurfaceKHR, createDisplayPlaneSurfaceKHR, createHeadlessSurfaceEXT, createIOSSurfaceMVK, createImagePipeSurfaceFUCHSIA, createMacOSSurfaceMVK, createMetalSurfaceEXT, createStreamDescriptorSurfaceGGP, createViSurfaceNN, createWaylandSurfaceKHR, createWin32SurfaceKHR, createXcbSurfaceKHR, createXlibSurfaceKHR, destroySurfaceKHR, getDeviceGroupSurfacePresentModesKHR, getPhysicalDevicePresentRectanglesKHR, getPhysicalDeviceSurfaceCapabilities2EXT, getPhysicalDeviceSurfaceCapabilitiesKHR, getPhysicalDeviceSurfaceFormatsKHR, getPhysicalDeviceSurfacePresentModesKHR, getPhysicalDeviceSurfaceSupportKHR

Constructors

SurfaceKHR Word64

data CAMetalLayer