Graphics.Vulkan.Extensions.VK_EXT_sample_locations

Documentation

withCStructAttachmentSampleLocationsEXT :: AttachmentSampleLocationsEXT -> (VkAttachmentSampleLocationsEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkAttachmentSampleLocationsEXT and marshal a AttachmentSampleLocationsEXT into it. The VkAttachmentSampleLocationsEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructAttachmentSampleLocationsEXT :: VkAttachmentSampleLocationsEXT -> IO AttachmentSampleLocationsEXT

A function to read a VkAttachmentSampleLocationsEXT and all additional structures in the pointer chain into a AttachmentSampleLocationsEXT.

data AttachmentSampleLocationsEXT

VkAttachmentSampleLocationsEXT - Structure specifying the sample locations state to use in the initial layout transition of attachments

Description

If the image referenced by the framebuffer attachment at index attachmentIndex was not created with VK_IMAGE_CREATE_SAMPLE_LOCATIONS_COMPATIBLE_DEPTH_BIT_EXT then the values specified in sampleLocationsInfo are ignored.

Valid Usage (Implicit)

See Also

VkRenderPassSampleLocationsBeginInfoEXT, VkSampleLocationsInfoEXT

Constructors

AttachmentSampleLocationsEXT

Fields

attachmentIndex :: Word32

sampleLocationsInfo :: SampleLocationsInfoEXT

withCStructMultisamplePropertiesEXT :: MultisamplePropertiesEXT -> (VkMultisamplePropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkMultisamplePropertiesEXT and marshal a MultisamplePropertiesEXT into it. The VkMultisamplePropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructMultisamplePropertiesEXT :: VkMultisamplePropertiesEXT -> IO MultisamplePropertiesEXT

A function to read a VkMultisamplePropertiesEXT and all additional structures in the pointer chain into a MultisamplePropertiesEXT.

data MultisamplePropertiesEXT

VkMultisamplePropertiesEXT - Structure returning information about sample count specific additional multisampling capabilities

Valid Usage (Implicit)

See Also

VkExtent2D, VkStructureType, vkGetPhysicalDeviceMultisamplePropertiesEXT

Constructors

MultisamplePropertiesEXT

Fields

next :: Maybe SomeVkStruct

maxSampleLocationGridSize :: Extent2D

withCStructPhysicalDeviceSampleLocationsPropertiesEXT :: PhysicalDeviceSampleLocationsPropertiesEXT -> (VkPhysicalDeviceSampleLocationsPropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceSampleLocationsPropertiesEXT and marshal a PhysicalDeviceSampleLocationsPropertiesEXT into it. The VkPhysicalDeviceSampleLocationsPropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceSampleLocationsPropertiesEXT :: VkPhysicalDeviceSampleLocationsPropertiesEXT -> IO PhysicalDeviceSampleLocationsPropertiesEXT

A function to read a VkPhysicalDeviceSampleLocationsPropertiesEXT and all additional structures in the pointer chain into a PhysicalDeviceSampleLocationsPropertiesEXT.

data PhysicalDeviceSampleLocationsPropertiesEXT

VkPhysicalDeviceSampleLocationsPropertiesEXT - Structure describing sample location limits that can be supported by an implementation

Members

The members of the VkPhysicalDeviceSampleLocationsPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceSampleLocationsPropertiesEXT structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkBool32, VkExtent2D, VkSampleCountFlags, VkStructureType

Constructors

PhysicalDeviceSampleLocationsPropertiesEXT

Fields

next :: Maybe SomeVkStruct

sampleLocationSampleCounts :: SampleCountFlags

maxSampleLocationGridSize :: Extent2D

sampleLocationCoordinateRange :: (CFloat, CFloat)

sampleLocationSubPixelBits :: Word32

variableSampleLocations :: Bool

withCStructPipelineSampleLocationsStateCreateInfoEXT :: PipelineSampleLocationsStateCreateInfoEXT -> (VkPipelineSampleLocationsStateCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPipelineSampleLocationsStateCreateInfoEXT and marshal a PipelineSampleLocationsStateCreateInfoEXT into it. The VkPipelineSampleLocationsStateCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPipelineSampleLocationsStateCreateInfoEXT :: VkPipelineSampleLocationsStateCreateInfoEXT -> IO PipelineSampleLocationsStateCreateInfoEXT

A function to read a VkPipelineSampleLocationsStateCreateInfoEXT and all additional structures in the pointer chain into a PipelineSampleLocationsStateCreateInfoEXT.

data PipelineSampleLocationsStateCreateInfoEXT

VkPipelineSampleLocationsStateCreateInfoEXT - Structure specifying sample locations for a pipeline

Valid Usage (Implicit)

See Also

VkBool32, VkSampleLocationsInfoEXT, VkStructureType

Constructors

PipelineSampleLocationsStateCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

sampleLocationsEnable :: Bool

sampleLocationsInfo :: SampleLocationsInfoEXT

withCStructRenderPassSampleLocationsBeginInfoEXT :: RenderPassSampleLocationsBeginInfoEXT -> (VkRenderPassSampleLocationsBeginInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkRenderPassSampleLocationsBeginInfoEXT and marshal a RenderPassSampleLocationsBeginInfoEXT into it. The VkRenderPassSampleLocationsBeginInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructRenderPassSampleLocationsBeginInfoEXT :: VkRenderPassSampleLocationsBeginInfoEXT -> IO RenderPassSampleLocationsBeginInfoEXT

A function to read a VkRenderPassSampleLocationsBeginInfoEXT and all additional structures in the pointer chain into a RenderPassSampleLocationsBeginInfoEXT.

data RenderPassSampleLocationsBeginInfoEXT

VkRenderPassSampleLocationsBeginInfoEXT - Structure specifying sample locations to use for the layout transition of custom sample locations compatible depth/stencil attachments

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_RENDER_PASS_SAMPLE_LOCATIONS_BEGIN_INFO_EXT

If attachmentInitialSampleLocationsCount is not 0, pAttachmentInitialSampleLocations must be a valid pointer to an array of attachmentInitialSampleLocationsCount valid VkAttachmentSampleLocationsEXT structures

If postSubpassSampleLocationsCount is not 0, pPostSubpassSampleLocations must be a valid pointer to an array of postSubpassSampleLocationsCount valid VkSubpassSampleLocationsEXT structures

See Also

VkAttachmentSampleLocationsEXT, VkStructureType, VkSubpassSampleLocationsEXT

Constructors

RenderPassSampleLocationsBeginInfoEXT

Fields

next :: Maybe SomeVkStruct

attachmentInitialSampleLocations :: Vector AttachmentSampleLocationsEXT

postSubpassSampleLocations :: Vector SubpassSampleLocationsEXT

withCStructSampleLocationEXT :: SampleLocationEXT -> (VkSampleLocationEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkSampleLocationEXT and marshal a SampleLocationEXT into it. The VkSampleLocationEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructSampleLocationEXT :: VkSampleLocationEXT -> IO SampleLocationEXT

A function to read a VkSampleLocationEXT and all additional structures in the pointer chain into a SampleLocationEXT.

data SampleLocationEXT

VkSampleLocationEXT - Structure specifying the coordinates of a sample location

Description

The domain space of the sample location coordinates has an upper-left origin within the pixel in framebuffer space.

The values specified in a VkSampleLocationEXT structure are always clamped to the implementation-dependent sample location coordinate range [sampleLocationCoordinateRange[0],sampleLocationCoordinateRange[1]] that can be queried by chaining the VkPhysicalDeviceSampleLocationsPropertiesEXT structure to the pNext chain of VkPhysicalDeviceProperties2.

See Also

VkSampleLocationsInfoEXT

Constructors

SampleLocationEXT

Fields

x :: CFloat

y :: CFloat

withCStructSampleLocationsInfoEXT :: SampleLocationsInfoEXT -> (VkSampleLocationsInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkSampleLocationsInfoEXT and marshal a SampleLocationsInfoEXT into it. The VkSampleLocationsInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructSampleLocationsInfoEXT :: VkSampleLocationsInfoEXT -> IO SampleLocationsInfoEXT

A function to read a VkSampleLocationsInfoEXT and all additional structures in the pointer chain into a SampleLocationsInfoEXT.

data SampleLocationsInfoEXT

VkSampleLocationsInfoEXT - Structure specifying a set of sample locations

Description

This structure can be used either to specify the sample locations to be used for rendering or to specify the set of sample locations an image subresource has been last rendered with for the purposes of layout transitions of depth/stencil images created with VK_IMAGE_CREATE_SAMPLE_LOCATIONS_COMPATIBLE_DEPTH_BIT_EXT.

The sample locations in pSampleLocations specify sampleLocationsPerPixel number of sample locations for each pixel in the grid of the size specified in sampleLocationGridSize. The sample location for sample i at the pixel grid location (x,y) is taken from pSampleLocations[(x + y * sampleLocationGridSize.width) * sampleLocationsPerPixel + i].

Valid Usage

sampleLocationsPerPixel must be a bit value that is set in VkPhysicalDeviceSampleLocationsPropertiesEXT::sampleLocationSampleCounts

sampleLocationsCount must equal sampleLocationsPerPixel × sampleLocationGridSize.width × sampleLocationGridSize.height

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_SAMPLE_LOCATIONS_INFO_EXT

If sampleLocationsPerPixel is not 0, sampleLocationsPerPixel must be a valid VkSampleCountFlagBits value

If sampleLocationsCount is not 0, pSampleLocations must be a valid pointer to an array of sampleLocationsCount VkSampleLocationEXT structures

See Also

VkAttachmentSampleLocationsEXT, VkExtent2D, VkPipelineSampleLocationsStateCreateInfoEXT, VkSampleCountFlagBits, VkSampleLocationEXT, VkStructureType, VkSubpassSampleLocationsEXT, vkCmdSetSampleLocationsEXT

Constructors

SampleLocationsInfoEXT

Fields

next :: Maybe SomeVkStruct

sampleLocationsPerPixel :: SampleCountFlagBits

sampleLocationGridSize :: Extent2D

sampleLocations :: Vector SampleLocationEXT

withCStructSubpassSampleLocationsEXT :: SubpassSampleLocationsEXT -> (VkSubpassSampleLocationsEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkSubpassSampleLocationsEXT and marshal a SubpassSampleLocationsEXT into it. The VkSubpassSampleLocationsEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructSubpassSampleLocationsEXT :: VkSubpassSampleLocationsEXT -> IO SubpassSampleLocationsEXT

A function to read a VkSubpassSampleLocationsEXT and all additional structures in the pointer chain into a SubpassSampleLocationsEXT.

data SubpassSampleLocationsEXT

VkSubpassSampleLocationsEXT - Structure specifying the sample locations state to use for layout transitions of attachments performed after a given subpass

Description

If the image referenced by the depth/stencil attachment used in the subpass identified by subpassIndex was not created with VK_IMAGE_CREATE_SAMPLE_LOCATIONS_COMPATIBLE_DEPTH_BIT_EXT or if the subpass does not use a depth/stencil attachment, and VkPhysicalDeviceSampleLocationsPropertiesEXT::variableSampleLocations is VK_TRUE then the values specified in sampleLocationsInfo are ignored.

Valid Usage (Implicit)

See Also

VkRenderPassSampleLocationsBeginInfoEXT, VkSampleLocationsInfoEXT

Constructors

SubpassSampleLocationsEXT

Fields

subpassIndex :: Word32

sampleLocationsInfo :: SampleLocationsInfoEXT

cmdSetSampleLocationsEXT :: CommandBuffer -> SampleLocationsInfoEXT -> IO ()

vkCmdSetSampleLocationsEXT - Set the dynamic sample locations state

Parameters

commandBuffer is the command buffer into which the command will be recorded.

pSampleLocationsInfo is the sample locations state to set.

Valid Usage

The bound graphics pipeline must have been created with the VK_DYNAMIC_STATE_SAMPLE_LOCATIONS_EXT dynamic state enabled

The sampleLocationsPerPixel member of pSampleLocationsInfo must equal the rasterizationSamples member of the VkPipelineMultisampleStateCreateInfo structure the bound graphics pipeline has been created with

If VkPhysicalDeviceSampleLocationsPropertiesEXT::variableSampleLocations is VK_FALSE then the current render pass must have been begun by specifying a VkRenderPassSampleLocationsBeginInfoEXT structure whose pPostSubpassSampleLocations member contains an element with a subpassIndex matching the current subpass index and the sampleLocationsInfo member of that element must match the sample locations state pointed to by pSampleLocationsInfo

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pSampleLocationsInfo must be a valid pointer to a valid VkSampleLocationsInfoEXT structure

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkSampleLocationsInfoEXT

getPhysicalDeviceMultisamplePropertiesEXT :: PhysicalDevice -> SampleCountFlagBits -> IO MultisamplePropertiesEXT

vkGetPhysicalDeviceMultisamplePropertiesEXT - Report sample count specific multisampling capabilities of a physical device

Parameters

physicalDevice is the physical device from which to query the additional multisampling capabilities.

samples is the sample count to query the capabilities for.

pMultisampleProperties is a pointer to a structure of type VkMultisamplePropertiesEXT, in which information about the additional multisampling capabilities specific to the sample count is returned.

Valid Usage (Implicit)

See Also

VkMultisamplePropertiesEXT, VkPhysicalDevice, VkSampleCountFlagBits

pattern EXT_SAMPLE_LOCATIONS_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_SAMPLE_LOCATIONS_SPEC_VERSION :: Integral a => a

pattern IMAGE_CREATE_SAMPLE_LOCATIONS_COMPATIBLE_DEPTH_BIT_EXT :: a ~ ImageCreateFlagBits => a

pattern STRUCTURE_TYPE_SAMPLE_LOCATIONS_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_RENDER_PASS_SAMPLE_LOCATIONS_BEGIN_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PIPELINE_SAMPLE_LOCATIONS_STATE_CREATE_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SAMPLE_LOCATIONS_PROPERTIES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_MULTISAMPLE_PROPERTIES_EXT :: a ~ StructureType => a

pattern DYNAMIC_STATE_SAMPLE_LOCATIONS_EXT :: a ~ DynamicState => a