Graphics.Vulkan.Extensions.VK_EXT_sampler_filter_minmax

Documentation

withCStructPhysicalDeviceSamplerFilterMinmaxPropertiesEXT :: PhysicalDeviceSamplerFilterMinmaxPropertiesEXT -> (VkPhysicalDeviceSamplerFilterMinmaxPropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceSamplerFilterMinmaxPropertiesEXT and marshal a PhysicalDeviceSamplerFilterMinmaxPropertiesEXT into it. The VkPhysicalDeviceSamplerFilterMinmaxPropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceSamplerFilterMinmaxPropertiesEXT :: VkPhysicalDeviceSamplerFilterMinmaxPropertiesEXT -> IO PhysicalDeviceSamplerFilterMinmaxPropertiesEXT

A function to read a VkPhysicalDeviceSamplerFilterMinmaxPropertiesEXT and all additional structures in the pointer chain into a PhysicalDeviceSamplerFilterMinmaxPropertiesEXT.

data PhysicalDeviceSamplerFilterMinmaxPropertiesEXT

VkPhysicalDeviceSamplerFilterMinmaxPropertiesEXT - Structure describing sampler filter minmax limits that can be supported by an implementation

Members

The members of the VkPhysicalDeviceSamplerFilterMinmaxPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceSamplerFilterMinmaxPropertiesEXT structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

If filterMinmaxSingleComponentFormats is VK_TRUE, the following formats must support the VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_MINMAX_BIT_EXT feature with VK_IMAGE_TILING_OPTIMAL, if they support VK_FORMAT_FEATURE_SAMPLED_IMAGE_BIT.

If the format is a depth/stencil format, this bit only specifies that the depth aspect (not the stencil aspect) of an image of this format supports min/max filtering, and that min/max filtering of the depth aspect is supported when depth compare is disabled in the sampler.

If filterMinmaxImageComponentMapping is VK_FALSE the component mapping of the image view used with min/max filtering must have been created with the r component set to VK_COMPONENT_SWIZZLE_IDENTITY. Only the r component of the sampled image value is defined and the other component values are undefined. If filterMinmaxImageComponentMapping is VK_TRUE this restriction does not apply and image component mapping works as normal.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceSamplerFilterMinmaxPropertiesEXT

Fields

next :: Maybe SomeVkStruct

filterMinmaxSingleComponentFormats :: Bool

filterMinmaxImageComponentMapping :: Bool

withCStructSamplerReductionModeCreateInfoEXT :: SamplerReductionModeCreateInfoEXT -> (VkSamplerReductionModeCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkSamplerReductionModeCreateInfoEXT and marshal a SamplerReductionModeCreateInfoEXT into it. The VkSamplerReductionModeCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructSamplerReductionModeCreateInfoEXT :: VkSamplerReductionModeCreateInfoEXT -> IO SamplerReductionModeCreateInfoEXT

A function to read a VkSamplerReductionModeCreateInfoEXT and all additional structures in the pointer chain into a SamplerReductionModeCreateInfoEXT.

data SamplerReductionModeCreateInfoEXT

VkSamplerReductionModeCreateInfoEXT - Structure specifying sampler reduction mode

Description

If this structure is not present, reductionMode is considered to be VK_SAMPLER_REDUCTION_MODE_WEIGHTED_AVERAGE_EXT.

Valid Usage (Implicit)

See Also

VkSamplerReductionModeEXT, VkStructureType

Constructors

SamplerReductionModeCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

reductionMode :: SamplerReductionModeEXT

type SamplerReductionModeEXT = VkSamplerReductionModeEXT

VkSamplerReductionModeEXT - Specify reduction mode for texture filtering

See Also

VkSamplerReductionModeCreateInfoEXT

pattern SAMPLER_REDUCTION_MODE_WEIGHTED_AVERAGE_EXT :: a ~ SamplerReductionModeEXT => a

VK_SAMPLER_REDUCTION_MODE_WEIGHTED_AVERAGE_EXT specifies that texel values are combined by computing a weighted average of values in the footprint, using weights as specified in the image operations chapter.

pattern SAMPLER_REDUCTION_MODE_MIN_EXT :: a ~ SamplerReductionModeEXT => a

VK_SAMPLER_REDUCTION_MODE_MIN_EXT specifies that texel values are combined by taking the component-wise minimum of values in the footprint with non-zero weights.

pattern SAMPLER_REDUCTION_MODE_MAX_EXT :: a ~ SamplerReductionModeEXT => a

VK_SAMPLER_REDUCTION_MODE_MAX_EXT specifies that texel values are combined by taking the component-wise maximum of values in the footprint with non-zero weights.

pattern EXT_SAMPLER_FILTER_MINMAX_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_SAMPLER_FILTER_MINMAX_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SAMPLER_FILTER_MINMAX_PROPERTIES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SAMPLER_REDUCTION_MODE_CREATE_INFO_EXT :: a ~ StructureType => a

pattern FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_MINMAX_BIT_EXT :: a ~ FormatFeatureFlagBits => a