Graphics.Vulkan.Extensions.VK_EXT_separate_stencil_usage

Documentation

withCStructImageStencilUsageCreateInfoEXT :: ImageStencilUsageCreateInfoEXT -> (VkImageStencilUsageCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkImageStencilUsageCreateInfoEXT and marshal a ImageStencilUsageCreateInfoEXT into it. The VkImageStencilUsageCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructImageStencilUsageCreateInfoEXT :: VkImageStencilUsageCreateInfoEXT -> IO ImageStencilUsageCreateInfoEXT

A function to read a VkImageStencilUsageCreateInfoEXT and all additional structures in the pointer chain into a ImageStencilUsageCreateInfoEXT.

data ImageStencilUsageCreateInfoEXT

VkImageStencilUsageCreateInfoEXT - Specify separate usage flags for the stencil aspect of a depth-stencil image

Description

This structure specifies image usages which only apply to the stencil aspect of a depth/stencil format image. When this structure is included in the pNext chain of VkImageCreateInfo, the stencil aspect of the image must only be used as specified by stencilUsage. When this structure is not included in the pNext chain of VkImageCreateInfo, the stencil aspect of an image must only be used as specified VkImageCreateInfo::usage. Use of other aspects of an image are unaffected by this structure.

This structure can also be included in the pNext chain of VkPhysicalDeviceImageFormatInfo2 to query additional capabilities specific to image creation parameter combinations including a separate set of usage flags for the stencil aspect of the image using vkGetPhysicalDeviceImageFormatProperties2. When this structure is not present in the pNext chain of VkPhysicalDeviceImageFormatInfo2 then the implicit value of stencilUsage matches that of VkPhysicalDeviceImageFormatInfo2::usage.

Valid Usage

If stencilUsage includes VK_IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT, then bits other than VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT, and VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT must not be set

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMAGE_STENCIL_USAGE_CREATE_INFO_EXT

stencilUsage must be a valid combination of VkImageUsageFlagBits values

stencilUsage must not be 0

See Also

VkImageUsageFlags, VkStructureType

Constructors

ImageStencilUsageCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

stencilUsage :: ImageUsageFlags

pattern EXT_SEPARATE_STENCIL_USAGE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_SEPARATE_STENCIL_USAGE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_IMAGE_STENCIL_USAGE_CREATE_INFO_EXT :: a ~ StructureType => a