Graphics.Vulkan.Extensions.VK_EXT_subgroup_size_control

Documentation

data PhysicalDeviceSubgroupSizeControlFeaturesEXT

VkPhysicalDeviceSubgroupSizeControlFeaturesEXT - Structure describing the subgroup size control features that can be supported by an implementation

Members

The members of the PhysicalDeviceSubgroupSizeControlFeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceSubgroupSizeControlFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether the feature is supported. PhysicalDeviceSubgroupSizeControlFeaturesEXT can also be included in the pNext chain of DeviceCreateInfo to enable the feature.

Note

The PhysicalDeviceSubgroupSizeControlFeaturesEXT structure was added in version 2 of the https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#VK_EXT_subgroup_size_control extension. Version 1 implementations of this extension will not fill out the features structure but applications may assume that both subgroupSizeControl and computeFullSubgroups are supported if the extension is supported. (See also the Feature Requirements section.) Applications are advised to add a PhysicalDeviceSubgroupSizeControlFeaturesEXT structure to the pNext chain of DeviceCreateInfo to enable the features regardless of the version of the extension supported by the implementation. If the implementation only supports version 1, it will safely ignore the PhysicalDeviceSubgroupSizeControlFeaturesEXT structure.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceSubgroupSizeControlFeaturesEXT

Fields

subgroupSizeControl :: Bool

subgroupSizeControl indicates whether the implementation supports controlling shader subgroup sizes via the PIPELINE_SHADER_STAGE_CREATE_ALLOW_VARYING_SUBGROUP_SIZE_BIT_EXT flag and the PipelineShaderStageRequiredSubgroupSizeCreateInfoEXT structure.

computeFullSubgroups :: Bool

computeFullSubgroups indicates whether the implementation supports requiring full subgroups in compute shaders via the PIPELINE_SHADER_STAGE_CREATE_REQUIRE_FULL_SUBGROUPS_BIT_EXT flag.

data PhysicalDeviceSubgroupSizeControlPropertiesEXT

VkPhysicalDeviceSubgroupSizeControlPropertiesEXT - Structure describing the control subgroup size properties of an implementation

Members

The members of the PhysicalDeviceSubgroupSizeControlPropertiesEXT structure describe the following properties:

Description

If the PhysicalDeviceSubgroupSizeControlPropertiesEXT structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

If PhysicalDeviceSubgroupProperties::supportedOperations includes https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#features-subgroup-quad, minSubgroupSize must be greater than or equal to 4.

Valid Usage (Implicit)

See Also

ShaderStageFlags, StructureType

Constructors

PhysicalDeviceSubgroupSizeControlPropertiesEXT

Fields

minSubgroupSize :: Word32

minSubgroupSize is the minimum subgroup size supported by this device. minSubgroupSize is at least one if any of the physical device’s queues support QUEUE_GRAPHICS_BIT or QUEUE_COMPUTE_BIT. minSubgroupSize is a power-of-two. minSubgroupSize is less than or equal to maxSubgroupSize. minSubgroupSize is less than or equal to subgroupSize.

maxSubgroupSize :: Word32

maxSubgroupSize is the maximum subgroup size supported by this device. maxSubgroupSize is at least one if any of the physical device’s queues support QUEUE_GRAPHICS_BIT or QUEUE_COMPUTE_BIT. maxSubgroupSize is a power-of-two. maxSubgroupSize is greater than or equal to minSubgroupSize. maxSubgroupSize is greater than or equal to subgroupSize.

maxComputeWorkgroupSubgroups :: Word32

maxComputeWorkgroupSubgroups is the maximum number of subgroups supported by the implementation within a workgroup.

requiredSubgroupSizeStages :: ShaderStageFlags

requiredSubgroupSizeStages is a bitfield of what shader stages support having a required subgroup size specified.

data PipelineShaderStageRequiredSubgroupSizeCreateInfoEXT

VkPipelineShaderStageRequiredSubgroupSizeCreateInfoEXT - Structure specifying the required subgroup size of a newly created pipeline shader stage

Valid Usage

Description

If a PipelineShaderStageRequiredSubgroupSizeCreateInfoEXT structure is included in the pNext chain of PipelineShaderStageCreateInfo, it specifies that the pipeline shader stage being compiled has a required subgroup size.

Valid Usage (Implicit)

See Also

StructureType

Constructors

PipelineShaderStageRequiredSubgroupSizeCreateInfoEXT

Fields

requiredSubgroupSize :: Word32

requiredSubgroupSize must be less than or equal to maxSubgroupSize.

type EXT_SUBGROUP_SIZE_CONTROL_SPEC_VERSION = 2

pattern EXT_SUBGROUP_SIZE_CONTROL_SPEC_VERSION :: forall a. Integral a => a

type EXT_SUBGROUP_SIZE_CONTROL_EXTENSION_NAME = "VK_EXT_subgroup_size_control"

pattern EXT_SUBGROUP_SIZE_CONTROL_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a