Graphics.Vulkan.Extensions.VK_EXT_texture_compression_astc_hdr

Documentation

data PhysicalDeviceTextureCompressionASTCHDRFeaturesEXT

VkPhysicalDeviceTextureCompressionASTCHDRFeaturesEXT - Structure describing ASTC HDR features that can be supported by an implementation

Members

The members of the PhysicalDeviceTextureCompressionASTCHDRFeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceTextureCompressionASTCHDRFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether each feature is supported. PhysicalDeviceTextureCompressionASTCHDRFeaturesEXT can also be included in the pNext chain of createDevice to enable features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceTextureCompressionASTCHDRFeaturesEXT

Fields

textureCompressionASTC_HDR :: Bool

textureCompressionASTC_HDR indicates whether all of the ASTC HDR compressed texture formats are supported. If this feature is enabled, then the FORMAT_FEATURE_SAMPLED_IMAGE_BIT, FORMAT_FEATURE_BLIT_SRC_BIT and FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_LINEAR_BIT features must be supported in optimalTilingFeatures for the following formats:

FORMAT_ASTC_4x4_SFLOAT_BLOCK_EXT

FORMAT_ASTC_5x4_SFLOAT_BLOCK_EXT

FORMAT_ASTC_5x5_SFLOAT_BLOCK_EXT

FORMAT_ASTC_6x5_SFLOAT_BLOCK_EXT

FORMAT_ASTC_6x6_SFLOAT_BLOCK_EXT

FORMAT_ASTC_8x5_SFLOAT_BLOCK_EXT

FORMAT_ASTC_8x6_SFLOAT_BLOCK_EXT

FORMAT_ASTC_8x8_SFLOAT_BLOCK_EXT

FORMAT_ASTC_10x5_SFLOAT_BLOCK_EXT

FORMAT_ASTC_10x6_SFLOAT_BLOCK_EXT

FORMAT_ASTC_10x8_SFLOAT_BLOCK_EXT

FORMAT_ASTC_10x10_SFLOAT_BLOCK_EXT

FORMAT_ASTC_12x10_SFLOAT_BLOCK_EXT

FORMAT_ASTC_12x12_SFLOAT_BLOCK_EXT

To query for additional properties, or if the feature is not enabled, getPhysicalDeviceFormatProperties and getPhysicalDeviceImageFormatProperties can be used to check for supported properties of individual formats as normal.

type EXT_TEXTURE_COMPRESSION_ASTC_HDR_SPEC_VERSION = 1

pattern EXT_TEXTURE_COMPRESSION_ASTC_HDR_SPEC_VERSION :: forall a. Integral a => a

type EXT_TEXTURE_COMPRESSION_ASTC_HDR_EXTENSION_NAME = "VK_EXT_texture_compression_astc_hdr"

pattern EXT_TEXTURE_COMPRESSION_ASTC_HDR_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a