Graphics.Vulkan.Extensions.VK_EXT_transform_feedback

Documentation

withCStructPhysicalDeviceTransformFeedbackFeaturesEXT :: PhysicalDeviceTransformFeedbackFeaturesEXT -> (VkPhysicalDeviceTransformFeedbackFeaturesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceTransformFeedbackFeaturesEXT and marshal a PhysicalDeviceTransformFeedbackFeaturesEXT into it. The VkPhysicalDeviceTransformFeedbackFeaturesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceTransformFeedbackFeaturesEXT :: VkPhysicalDeviceTransformFeedbackFeaturesEXT -> IO PhysicalDeviceTransformFeedbackFeaturesEXT

A function to read a VkPhysicalDeviceTransformFeedbackFeaturesEXT and all additional structures in the pointer chain into a PhysicalDeviceTransformFeedbackFeaturesEXT.

data PhysicalDeviceTransformFeedbackFeaturesEXT

VkPhysicalDeviceTransformFeedbackFeaturesEXT - Structure describing transform feedback features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceTransformFeedbackFeaturesEXT structure describe the following features:

Description

If the VkPhysicalDeviceTransformFeedbackFeaturesEXT structure is included in the pNext chain of VkPhysicalDeviceFeatures2KHR, it is filled with values indicating whether each feature is supported. VkPhysicalDeviceTransformFeedbackFeaturesEXT can also be used in the pNext chain of VkDeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceTransformFeedbackFeaturesEXT

Fields

next :: Maybe SomeVkStruct

transformFeedback :: Bool

geometryStreams :: Bool

withCStructPhysicalDeviceTransformFeedbackPropertiesEXT :: PhysicalDeviceTransformFeedbackPropertiesEXT -> (VkPhysicalDeviceTransformFeedbackPropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceTransformFeedbackPropertiesEXT and marshal a PhysicalDeviceTransformFeedbackPropertiesEXT into it. The VkPhysicalDeviceTransformFeedbackPropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceTransformFeedbackPropertiesEXT :: VkPhysicalDeviceTransformFeedbackPropertiesEXT -> IO PhysicalDeviceTransformFeedbackPropertiesEXT

A function to read a VkPhysicalDeviceTransformFeedbackPropertiesEXT and all additional structures in the pointer chain into a PhysicalDeviceTransformFeedbackPropertiesEXT.

data PhysicalDeviceTransformFeedbackPropertiesEXT

VkPhysicalDeviceTransformFeedbackPropertiesEXT - Structure describing transform feedback properties that can be supported by an implementation

Members

The members of the VkPhysicalDeviceTransformFeedbackPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceTransformFeedbackPropertiesEXT structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits and properties.

Valid Usage (Implicit)

See Also

VkBool32, VkDeviceSize, VkStructureType

Constructors

PhysicalDeviceTransformFeedbackPropertiesEXT

Fields

next :: Maybe SomeVkStruct

maxTransformFeedbackStreams :: Word32

maxTransformFeedbackBuffers :: Word32

maxTransformFeedbackBufferSize :: DeviceSize

maxTransformFeedbackStreamDataSize :: Word32

maxTransformFeedbackBufferDataSize :: Word32

maxTransformFeedbackBufferDataStride :: Word32

transformFeedbackQueries :: Bool

transformFeedbackStreamsLinesTriangles :: Bool

transformFeedbackRasterizationStreamSelect :: Bool

transformFeedbackDraw :: Bool

type PipelineRasterizationStateStreamCreateFlagsEXT = VkPipelineRasterizationStateStreamCreateFlagsEXT

VkPipelineRasterizationStateStreamCreateFlagsEXT - Reserved for future use

Description

VkPipelineRasterizationStateStreamCreateFlagsEXT is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkPipelineRasterizationStateStreamCreateInfoEXT

withCStructPipelineRasterizationStateStreamCreateInfoEXT :: PipelineRasterizationStateStreamCreateInfoEXT -> (VkPipelineRasterizationStateStreamCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPipelineRasterizationStateStreamCreateInfoEXT and marshal a PipelineRasterizationStateStreamCreateInfoEXT into it. The VkPipelineRasterizationStateStreamCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPipelineRasterizationStateStreamCreateInfoEXT :: VkPipelineRasterizationStateStreamCreateInfoEXT -> IO PipelineRasterizationStateStreamCreateInfoEXT

A function to read a VkPipelineRasterizationStateStreamCreateInfoEXT and all additional structures in the pointer chain into a PipelineRasterizationStateStreamCreateInfoEXT.

data PipelineRasterizationStateStreamCreateInfoEXT

VkPipelineRasterizationStateStreamCreateInfoEXT - Structure defining the geometry stream used for rasterization

Description

If this structure is not present, rasterizationStream is assumed to be zero.

Valid Usage (Implicit)

See Also

VkPipelineRasterizationStateStreamCreateFlagsEXT, VkStructureType

Constructors

PipelineRasterizationStateStreamCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

flags :: PipelineRasterizationStateStreamCreateFlagsEXT

rasterizationStream :: Word32

cmdBeginQueryIndexedEXT :: CommandBuffer -> QueryPool -> Word32 -> QueryControlFlags -> Word32 -> IO ()

vkCmdBeginQueryIndexedEXT - Begin an indexed query

Parameters

commandBuffer is the command buffer into which this command will be recorded.

queryPool is the query pool that will manage the results of the query.

query is the query index within the query pool that will contain the results.

flags is a bitmask of VkQueryControlFlagBits specifying constraints on the types of queries that can be performed.

index is the query type specific index. When the query type is VK_QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT the index represents the vertex stream.

Description

The vkCmdBeginQueryIndexedEXT command operates the same as the vkCmdBeginQuery command, except that it also accepts a query type specific index parameter.

Valid Usage

queryPool must have been created with a queryType that differs from that of any queries that are active within commandBuffer

All queries used by the command must be unavailable

If the precise occlusion queries feature is not enabled, or the queryType used to create queryPool was not VK_QUERY_TYPE_OCCLUSION, flags must not contain VK_QUERY_CONTROL_PRECISE_BIT

query must be less than the number of queries in queryPool

If the queryType used to create queryPool was VK_QUERY_TYPE_OCCLUSION, the VkCommandPool that commandBuffer was allocated from must support graphics operations

If the queryType used to create queryPool was VK_QUERY_TYPE_PIPELINE_STATISTICS and any of the pipelineStatistics indicate graphics operations, the VkCommandPool that commandBuffer was allocated from must support graphics operations

If the queryType used to create queryPool was VK_QUERY_TYPE_PIPELINE_STATISTICS and any of the pipelineStatistics indicate compute operations, the VkCommandPool that commandBuffer was allocated from must support compute operations

If the queryType used to create queryPool was VK_QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT the VkCommandPool that commandBuffer was allocated from must support graphics operations

If the queryType used to create queryPool was VK_QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT the index parameter must be less than VkPhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackStreams

If the queryType used to create queryPool was not VK_QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT the index must be zero

If the queryType used to create queryPool was VK_QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT then VkPhysicalDeviceTransformFeedbackPropertiesEXT::transformFeedbackQueries must be supported

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

queryPool must be a valid VkQueryPool handle

flags must be a valid combination of VkQueryControlFlagBits values

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

Both of commandBuffer, and queryPool must have been created, allocated, or retrieved from the same VkDevice

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkQueryControlFlags, VkQueryPool

cmdBeginTransformFeedbackEXT :: CommandBuffer -> Word32 -> Vector Buffer -> Maybe (Vector DeviceSize) -> IO ()

vkCmdBeginTransformFeedbackEXT - Make transform feedback active in the command buffer

Parameters

commandBuffer is the command buffer into which the command is recorded.

firstCounterBuffer is the index of the first transform feedback buffer corresponding to pCounterBuffers[0] and pCounterBufferOffsets[0].

counterBufferCount is the size of the pCounterBuffers and pCounterBufferOffsets arrays.

pCounterBuffers is an optional array of buffer handles to the counter buffers which contain a 4 byte integer value representing the byte offset from the start of the corresponding transform feedback buffer from where to start capturing vertex data. If the byte offset stored to the counter buffer location was done using vkCmdEndTransformFeedbackEXT it can be used to resume transform feedback from the previous location. If pCounterBuffers is NULL, then transform feedback will start capturing vertex data to byte offset zero in all bound transform feedback buffers. For each element of pCounterBuffers that is VK_NULL_HANDLE, transform feedback will start capturing vertex data to byte zero in the corresponding bound transform feedback buffer.

pCounterBufferOffsets is an optional array of offsets within each of the pCounterBuffers where the counter values were previously written. The location in each counter buffer at these offsets must be large enough to contain 4 bytes of data. This data is the number of bytes captured by the previous transform feedback to this buffer. If pCounterBufferOffsets is NULL, then it is assumed the offsets are zero.

Description

The active transform feedback buffers will capture primitives emitted from the corresponding XfbBuffer in the bound graphics pipeline. Any XfbBuffer emitted that does not output to an active transform feedback buffer will not be captured.

Valid Usage

VkPhysicalDeviceTransformFeedbackFeaturesEXT::transformFeedback must be enabled

Transform feedback must not be active

firstCounterBuffer must be less than VkPhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackBuffers

The sum of firstCounterBuffer and counterBufferCount must be less than or equal to VkPhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackBuffers

If counterBufferCount is not 0, and pCounterBuffers is not NULL, pCounterBuffers must be a valid pointer to an array of counterBufferCount VkBuffer handles that are either valid or VK_NULL_HANDLE

For each buffer handle in the array, if it is not VK_NULL_HANDLE it must reference a buffer large enough to hold 4 bytes at the corresponding offset from the pCounterBufferOffsets array

If pCounterBuffer is NULL, then pCounterBufferOffsets must also be NULL

For each buffer handle in the pCounterBuffers array that is not VK_NULL_HANDLE it must have been created with a usage value containing VK_BUFFER_USAGE_TRANSFORM_FEEDBACK_COUNTER_BUFFER_BIT_EXT

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

If counterBufferCount is not 0, and pCounterBufferOffsets is not NULL, pCounterBufferOffsets must be a valid pointer to an array of counterBufferCount VkDeviceSize values

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics operations

This command must only be called inside of a render pass instance

Both of commandBuffer, and the elements of pCounterBuffers that are valid handles must have been created, allocated, or retrieved from the same VkDevice

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Inside          | Graphics        |                 |
| Secondary       |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkBuffer, VkCommandBuffer, VkDeviceSize

cmdBindTransformFeedbackBuffersEXT :: CommandBuffer -> Word32 -> Vector Buffer -> Vector DeviceSize -> Maybe (Vector DeviceSize) -> IO ()

vkCmdBindTransformFeedbackBuffersEXT - Bind transform feedback buffers to a command buffer

Parameters

commandBuffer is the command buffer into which the command is recorded.

firstBinding is the index of the first transform feedback binding whose state is updated by the command.

bindingCount is the number of transform feedback bindings whose state is updated by the command.

pBuffers is a pointer to an array of buffer handles.

pOffsets is a pointer to an array of buffer offsets.

pSizes is an optional array of buffer sizes, which specifies the maximum number of bytes to capture to the corresponding transform feedback buffer. If pSizes is NULL, or the value of the pSizes array element is VK_WHOLE_SIZE, then the maximum bytes captured will be the size of the corresponding buffer minus the buffer offset.

Description

The values taken from elements i of pBuffers, pOffsets and pSizes replace the current state for the transform feedback binding firstBinding + i, for i in [0, bindingCount). The transform feedback binding is updated to start at the offset indicated by pOffsets[i] from the start of the buffer pBuffers[i].

Valid Usage

VkPhysicalDeviceTransformFeedbackFeaturesEXT::transformFeedback must be enabled

firstBinding must be less than VkPhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackBuffers

The sum of firstBinding and bindingCount must be less than or equal to VkPhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackBuffers

All elements of pOffsets must be less than the size of the corresponding element in pBuffers

All elements of pOffsets must be a multiple of 4

All elements of pBuffers must have been created with the VK_BUFFER_USAGE_TRANSFORM_FEEDBACK_BUFFER_BIT_EXT flag

If the optional pSize array is specified, each element of pSizes must either be VK_WHOLE_SIZE, or be less than or equal to VkPhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackBufferSize

All elements of pSizes must be less than or equal to the size of the corresponding buffer in pBuffers

All elements of pOffsets plus pSizes, where the pSizes, element is not VK_WHOLE_SIZE, must be less than or equal to the size of the corresponding element in pBuffers

Each element of pBuffers that is non-sparse must be bound completely and contiguously to a single VkDeviceMemory object

Transform feedback must not be active when the vkCmdBindTransformFeedbackBuffersEXT command is recorded

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pBuffers must be a valid pointer to an array of bindingCount valid VkBuffer handles

pOffsets must be a valid pointer to an array of bindingCount VkDeviceSize values

If pSizes is not NULL, pSizes must be a valid pointer to an array of bindingCount VkDeviceSize values

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics operations

If pSizes is not NULL, bindingCount must be greater than 0

Both of commandBuffer, and the elements of pBuffers must have been created, allocated, or retrieved from the same VkDevice

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkBuffer, VkCommandBuffer, VkDeviceSize

cmdDrawIndirectByteCountEXT :: CommandBuffer -> Word32 -> Word32 -> Buffer -> DeviceSize -> Word32 -> Word32 -> IO ()

vkCmdDrawIndirectByteCountEXT - Draw primitives where the vertex count is derived from the counter byte value in the counter buffer

Parameters

commandBuffer is the command buffer into which the command is recorded.

instanceCount is the number of instances to draw.

firstInstance is the instance ID of the first instance to draw.

counterBuffer is the buffer handle from where the byte count is read.

counterBufferOffset is the offset into the buffer used to read the byte count, which is used to calculate the vertex count for this draw call.

counterOffset is subtracted from the byte count read from the counterBuffer at the counterBufferOffset

vertexStride is the stride in bytes between each element of the vertex data that is used to calculate the vertex count from the counter value. This value is typically the same value that was used in the graphics pipeline state when the transform feedback was captured as the XfbStride.

Description

When the command is executed, primitives are assembled in the same way as done with vkCmdDraw except the vertexCount is calculated based on the byte count read from counterBuffer at offset counterBufferOffset. The assembled primitives execute the bound graphics pipeline.

The effective vertexCount is calculated as follows:

const uint32_t * counterBufferPtr = (const uint8_t *)counterBuffer.address + counterBufferOffset;
vertexCount = floor(max(0, (*counterBufferPtr - counterOffset)) / vertexStride);

The effective firstVertex is zero.

Valid Usage

If a VkImageView is sampled with VK_FILTER_LINEAR as a result of this command, then the image view’s format features must contain VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_LINEAR_BIT

If a VkImageView is accessed using atomic operations as a result of this command, then the image view’s format features must contain VK_FORMAT_FEATURE_STORAGE_IMAGE_ATOMIC_BIT

For each set n that is statically used by the VkPipeline bound to the pipeline bind point used by this command, a descriptor set must have been bound to n at the same pipeline bind point, with a VkPipelineLayout that is compatible for set n, with the VkPipelineLayout used to create the current VkPipeline, as described in ???

For each push constant that is statically used by the VkPipeline bound to the pipeline bind point used by this command, a push constant value must have been set for the same pipeline bind point, with a VkPipelineLayout that is compatible for push constants, with the VkPipelineLayout used to create the current VkPipeline, as described in ???

Descriptors in each bound descriptor set, specified via vkCmdBindDescriptorSets, must be valid if they are statically used by the VkPipeline bound to the pipeline bind point used by this command

A valid pipeline must be bound to the pipeline bind point used by this command

If the VkPipeline object bound to the pipeline bind point used by this command requires any dynamic state, that state must have been set for commandBuffer

If the VkPipeline object bound to the pipeline bind point used by this command accesses a VkSampler object that uses unnormalized coordinates, that sampler must not be used to sample from any VkImage with a VkImageView of the type VK_IMAGE_VIEW_TYPE_3D, VK_IMAGE_VIEW_TYPE_CUBE, VK_IMAGE_VIEW_TYPE_1D_ARRAY, VK_IMAGE_VIEW_TYPE_2D_ARRAY or VK_IMAGE_VIEW_TYPE_CUBE_ARRAY, in any shader stage

If the VkPipeline object bound to the pipeline bind point used by this command accesses a VkSampler object that uses unnormalized coordinates, that sampler must not be used with any of the SPIR-V OpImageSample* or OpImageSparseSample* instructions with ImplicitLod, Dref or Proj in their name, in any shader stage

If the VkPipeline object bound to the pipeline bind point used by this command accesses a VkSampler object that uses unnormalized coordinates, that sampler must not be used with any of the SPIR-V OpImageSample* or OpImageSparseSample* instructions that includes a LOD bias or any offset values, in any shader stage

If the robust buffer access feature is not enabled, and if the VkPipeline object bound to the pipeline bind point used by this command accesses a uniform buffer, it must not access values outside of the range of the buffer as specified in the descriptor set bound to the same pipeline bind point

If the robust buffer access feature is not enabled, and if the VkPipeline object bound to the pipeline bind point used by this command accesses a storage buffer, it must not access values outside of the range of the buffer as specified in the descriptor set bound to the same pipeline bind point

The current render pass must be compatible with the renderPass member of the VkGraphicsPipelineCreateInfo structure specified when creating the VkPipeline bound to VK_PIPELINE_BIND_POINT_GRAPHICS.

The subpass index of the current render pass must be equal to the subpass member of the VkGraphicsPipelineCreateInfo structure specified when creating the VkPipeline bound to VK_PIPELINE_BIND_POINT_GRAPHICS.

Every input attachment used by the current subpass must be bound to the pipeline via a descriptor set

Image subresources used as attachments in the current render pass must not be accessed in any way other than as an attachment by this command.

All vertex input bindings accessed via vertex input variables declared in the vertex shader entry point’s interface must have valid buffers bound

For a given vertex buffer binding, any attribute data fetched must be entirely contained within the corresponding vertex buffer binding, as described in ???

VkPhysicalDeviceTransformFeedbackFeaturesEXT::transformFeedback must be enabled

The implementation must support VkPhysicalDeviceTransformFeedbackPropertiesEXT::transformFeedbackDraw

vertexStride must be greater than 0 and less than or equal to VkPhysicalDeviceLimits::maxTransformFeedbackBufferDataStride

counterBuffer must have been created with the VK_BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

counterBuffer must be a valid VkBuffer handle

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics operations

This command must only be called inside of a render pass instance

Both of commandBuffer, and counterBuffer must have been created, allocated, or retrieved from the same VkDevice

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Inside          | Graphics        | Graphics        |
| Secondary       |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkBuffer, VkCommandBuffer, VkDeviceSize

cmdEndQueryIndexedEXT :: CommandBuffer -> QueryPool -> Word32 -> Word32 -> IO ()

vkCmdEndQueryIndexedEXT - Ends a query

Parameters

commandBuffer is the command buffer into which this command will be recorded.

queryPool is the query pool that is managing the results of the query.

query is the query index within the query pool where the result is stored.

index is the query type specific index.

Description

The vkCmdEndQueryIndexedEXT command operates the same as the vkCmdEndQuery command, except that it also accepts a query type specific index parameter.

Valid Usage

All queries used by the command must be active

query must be less than the number of queries in queryPool

If the queryType used to create queryPool was VK_QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT the index parameter must be less than VkPhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackStreams

If the queryType used to create queryPool was not VK_QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT the index must be zero

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

queryPool must be a valid VkQueryPool handle

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

Both of commandBuffer, and queryPool must have been created, allocated, or retrieved from the same VkDevice

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkQueryPool

cmdEndTransformFeedbackEXT :: CommandBuffer -> Word32 -> Vector Buffer -> Maybe (Vector DeviceSize) -> IO ()

vkCmdEndTransformFeedbackEXT - Make transform feedback inactive in the command buffer

Parameters

commandBuffer is the command buffer into which the command is recorded.

firstCounterBuffer is the index of the first transform feedback buffer corresponding to pCounterBuffers[0] and pCounterBufferOffsets[0].

counterBufferCount is the size of the pCounterBuffers and pCounterBufferOffsets arrays.

pCounterBuffers is an optional array of buffer handles to the counter buffers used to record the current byte positions of each transform feedback buffer where the next vertex output data would be captured. This can be used by a subsequent vkCmdBeginTransformFeedbackEXT call to resume transform feedback capture from this position. It can also be used by vkCmdDrawIndirectByteCountEXT to determine the vertex count of the draw call.

pCounterBufferOffsets is an optional array of offsets within each of the pCounterBuffers where the counter values can be written. The location in each counter buffer at these offsets must be large enough to contain 4 bytes of data. The data stored at this location is the byte offset from the start of the transform feedback buffer binding where the next vertex data would be written. If pCounterBufferOffsets is NULL, then it is assumed the offsets are zero.

Valid Usage

VkPhysicalDeviceTransformFeedbackFeaturesEXT::transformFeedback must be enabled

Transform feedback must be active

firstCounterBuffer must be less than VkPhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackBuffers

The sum of firstCounterBuffer and counterBufferCount must be less than or equal to VkPhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackBuffers

If counterBufferCount is not 0, and pCounterBuffers is not NULL, pCounterBuffers must be a valid pointer to an array of counterBufferCount VkBuffer handles that are either valid or VK_NULL_HANDLE

For each buffer handle in the array, if it is not VK_NULL_HANDLE it must reference a buffer large enough to hold 4 bytes at the corresponding offset from the pCounterBufferOffsets array

If pCounterBuffer is NULL, then pCounterBufferOffsets must also be NULL

For each buffer handle in the pCounterBuffers array that is not VK_NULL_HANDLE it must have been created with a usage value containing VK_BUFFER_USAGE_TRANSFORM_FEEDBACK_COUNTER_BUFFER_BIT_EXT

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

If counterBufferCount is not 0, and pCounterBufferOffsets is not NULL, pCounterBufferOffsets must be a valid pointer to an array of counterBufferCount VkDeviceSize values

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics operations

This command must only be called inside of a render pass instance

Both of commandBuffer, and the elements of pCounterBuffers that are valid handles must have been created, allocated, or retrieved from the same VkDevice

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Inside          | Graphics        |                 |
| Secondary       |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkBuffer, VkCommandBuffer, VkDeviceSize

pattern EXT_TRANSFORM_FEEDBACK_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_TRANSFORM_FEEDBACK_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_TRANSFORM_FEEDBACK_FEATURES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_TRANSFORM_FEEDBACK_PROPERTIES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_STREAM_CREATE_INFO_EXT :: a ~ StructureType => a

pattern QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT :: a ~ QueryType => a

pattern BUFFER_USAGE_TRANSFORM_FEEDBACK_BUFFER_BIT_EXT :: a ~ BufferUsageFlagBits => a

pattern BUFFER_USAGE_TRANSFORM_FEEDBACK_COUNTER_BUFFER_BIT_EXT :: a ~ BufferUsageFlagBits => a

pattern ACCESS_TRANSFORM_FEEDBACK_WRITE_BIT_EXT :: a ~ AccessFlagBits => a

pattern ACCESS_TRANSFORM_FEEDBACK_COUNTER_READ_BIT_EXT :: a ~ AccessFlagBits => a

pattern ACCESS_TRANSFORM_FEEDBACK_COUNTER_WRITE_BIT_EXT :: a ~ AccessFlagBits => a

pattern PIPELINE_STAGE_TRANSFORM_FEEDBACK_BIT_EXT :: a ~ PipelineStageFlagBits => a