Graphics.Vulkan.Extensions.VK_EXT_validation_features

Documentation

type ValidationFeatureDisableEXT = VkValidationFeatureDisableEXT

VkValidationFeatureDisableEXT - Specify validation features to disable

See Also

VkValidationFeaturesEXT

pattern VALIDATION_FEATURE_DISABLE_ALL_EXT :: a ~ ValidationFeatureDisableEXT => a

VK_VALIDATION_FEATURE_DISABLE_ALL_EXT specifies that all validation checks are disabled.

pattern VALIDATION_FEATURE_DISABLE_SHADERS_EXT :: a ~ ValidationFeatureDisableEXT => a

VK_VALIDATION_FEATURE_DISABLE_SHADERS_EXT specifies that shader validation is disabled. This feature is enabled by default.

pattern VALIDATION_FEATURE_DISABLE_THREAD_SAFETY_EXT :: a ~ ValidationFeatureDisableEXT => a

VK_VALIDATION_FEATURE_DISABLE_THREAD_SAFETY_EXT specifies that thread safety validation is disabled. This feature is enabled by default.

pattern VALIDATION_FEATURE_DISABLE_API_PARAMETERS_EXT :: a ~ ValidationFeatureDisableEXT => a

VK_VALIDATION_FEATURE_DISABLE_API_PARAMETERS_EXT specifies that stateless parameter validation is disabled. This feature is enabled by default.

pattern VALIDATION_FEATURE_DISABLE_OBJECT_LIFETIMES_EXT :: a ~ ValidationFeatureDisableEXT => a

VK_VALIDATION_FEATURE_DISABLE_OBJECT_LIFETIMES_EXT specifies that object lifetime validation is disabled. This feature is enabled by default.

pattern VALIDATION_FEATURE_DISABLE_CORE_CHECKS_EXT :: a ~ ValidationFeatureDisableEXT => a

VK_VALIDATION_FEATURE_DISABLE_CORE_CHECKS_EXT specifies that core validation checks are disabled. This feature is enabled by default. If this feature is disabled, the shader validation and GPU-assisted validation features are also disabled.

pattern VALIDATION_FEATURE_DISABLE_UNIQUE_HANDLES_EXT :: a ~ ValidationFeatureDisableEXT => a

VK_VALIDATION_FEATURE_DISABLE_UNIQUE_HANDLES_EXT specifies that protection against duplicate non-dispatchable object handles is disabled. This feature is enabled by default.

type ValidationFeatureEnableEXT = VkValidationFeatureEnableEXT

VkValidationFeatureEnableEXT - Specify validation features to enable

See Also

VkValidationFeaturesEXT

pattern VALIDATION_FEATURE_ENABLE_GPU_ASSISTED_EXT :: a ~ ValidationFeatureEnableEXT => a

VK_VALIDATION_FEATURE_ENABLE_GPU_ASSISTED_EXT specifies that GPU-assisted validation is enabled. Activating this feature instruments shader programs to generate additional diagnostic data. This feature is disabled by default.

pattern VALIDATION_FEATURE_ENABLE_GPU_ASSISTED_RESERVE_BINDING_SLOT_EXT :: a ~ ValidationFeatureEnableEXT => a

VK_VALIDATION_FEATURE_ENABLE_GPU_ASSISTED_RESERVE_BINDING_SLOT_EXT specifies that the validation layers reserve a descriptor set binding slot for their own use. The layer reports a value for VkPhysicalDeviceLimits::maxBoundDescriptorSets that is one less than the value reported by the device. If the device supports the binding of only one descriptor set, the validation layer does not perform GPU-assisted validation. This feature is disabled by default. The GPU-assisted validation feature must be enabled in order to use this feature.

withCStructValidationFeaturesEXT :: ValidationFeaturesEXT -> (VkValidationFeaturesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkValidationFeaturesEXT and marshal a ValidationFeaturesEXT into it. The VkValidationFeaturesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructValidationFeaturesEXT :: VkValidationFeaturesEXT -> IO ValidationFeaturesEXT

A function to read a VkValidationFeaturesEXT and all additional structures in the pointer chain into a ValidationFeaturesEXT.

data ValidationFeaturesEXT

VkValidationFeaturesEXT - Specify validation features to enable or disable for a Vulkan instance

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_VALIDATION_FEATURES_EXT

If enabledValidationFeatureCount is not 0, pEnabledValidationFeatures must be a valid pointer to an array of enabledValidationFeatureCount valid VkValidationFeatureEnableEXT values

If disabledValidationFeatureCount is not 0, pDisabledValidationFeatures must be a valid pointer to an array of disabledValidationFeatureCount valid VkValidationFeatureDisableEXT values

See Also

VkStructureType, VkValidationFeatureDisableEXT, VkValidationFeatureEnableEXT

Constructors

ValidationFeaturesEXT

Fields

next :: Maybe SomeVkStruct

enabledValidationFeatures :: Vector ValidationFeatureEnableEXT

disabledValidationFeatures :: Vector ValidationFeatureDisableEXT

pattern EXT_VALIDATION_FEATURES_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_VALIDATION_FEATURES_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_VALIDATION_FEATURES_EXT :: a ~ StructureType => a