Graphics.Vulkan.Extensions.VK_EXT_validation_flags

Documentation

type ValidationCheckEXT = VkValidationCheckEXT

VkValidationCheckEXT - Specify validation checks to disable

See Also

VkValidationFlagsEXT

pattern VALIDATION_CHECK_ALL_EXT :: a ~ ValidationCheckEXT => a

VK_VALIDATION_CHECK_ALL_EXT specifies that all validation checks are disabled.

pattern VALIDATION_CHECK_SHADERS_EXT :: a ~ ValidationCheckEXT => a

VK_VALIDATION_CHECK_SHADERS_EXT specifies that shader validation is disabled.

withCStructValidationFlagsEXT :: ValidationFlagsEXT -> (VkValidationFlagsEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkValidationFlagsEXT and marshal a ValidationFlagsEXT into it. The VkValidationFlagsEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructValidationFlagsEXT :: VkValidationFlagsEXT -> IO ValidationFlagsEXT

A function to read a VkValidationFlagsEXT and all additional structures in the pointer chain into a ValidationFlagsEXT.

data ValidationFlagsEXT

VkValidationFlagsEXT - Specify validation checks to disable for a Vulkan instance

Valid Usage (Implicit)

See Also

VkStructureType, VkValidationCheckEXT

Constructors

ValidationFlagsEXT

Fields

next :: Maybe SomeVkStruct

disabledValidationChecks :: Vector ValidationCheckEXT

pattern EXT_VALIDATION_FLAGS_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_VALIDATION_FLAGS_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_VALIDATION_FLAGS_EXT :: a ~ StructureType => a