Graphics.Vulkan.Extensions.VK_EXT_vertex_attribute_divisor

Documentation

withCStructPhysicalDeviceVertexAttributeDivisorFeaturesEXT :: PhysicalDeviceVertexAttributeDivisorFeaturesEXT -> (VkPhysicalDeviceVertexAttributeDivisorFeaturesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceVertexAttributeDivisorFeaturesEXT and marshal a PhysicalDeviceVertexAttributeDivisorFeaturesEXT into it. The VkPhysicalDeviceVertexAttributeDivisorFeaturesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceVertexAttributeDivisorFeaturesEXT :: VkPhysicalDeviceVertexAttributeDivisorFeaturesEXT -> IO PhysicalDeviceVertexAttributeDivisorFeaturesEXT

A function to read a VkPhysicalDeviceVertexAttributeDivisorFeaturesEXT and all additional structures in the pointer chain into a PhysicalDeviceVertexAttributeDivisorFeaturesEXT.

data PhysicalDeviceVertexAttributeDivisorFeaturesEXT

VkPhysicalDeviceVertexAttributeDivisorFeaturesEXT - Structure describing if fetching of vertex attribute may be repeated for instanced rendering

Description

If the VkPhysicalDeviceVertexAttributeDivisorFeaturesEXT structure is included in the pNext chain of VkPhysicalDeviceFeatures2, it is filled with values indicating the implementation-dependent behavior. VkPhysicalDeviceVertexAttributeDivisorFeaturesEXT can also be used in pNext chain of VkDeviceCreateInfo to enable the feature.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceVertexAttributeDivisorFeaturesEXT

Fields

next :: Maybe SomeVkStruct

vertexAttributeInstanceRateDivisor :: Bool

vertexAttributeInstanceRateZeroDivisor :: Bool

withCStructPhysicalDeviceVertexAttributeDivisorPropertiesEXT :: PhysicalDeviceVertexAttributeDivisorPropertiesEXT -> (VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT and marshal a PhysicalDeviceVertexAttributeDivisorPropertiesEXT into it. The VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceVertexAttributeDivisorPropertiesEXT :: VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT -> IO PhysicalDeviceVertexAttributeDivisorPropertiesEXT

A function to read a VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT and all additional structures in the pointer chain into a PhysicalDeviceVertexAttributeDivisorPropertiesEXT.

data PhysicalDeviceVertexAttributeDivisorPropertiesEXT

VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT - Structure describing max value of vertex attribute divisor that can be supported by an implementation

Members

The members of the VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

PhysicalDeviceVertexAttributeDivisorPropertiesEXT

Fields

next :: Maybe SomeVkStruct

maxVertexAttribDivisor :: Word32

withCStructPipelineVertexInputDivisorStateCreateInfoEXT :: PipelineVertexInputDivisorStateCreateInfoEXT -> (VkPipelineVertexInputDivisorStateCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkPipelineVertexInputDivisorStateCreateInfoEXT and marshal a PipelineVertexInputDivisorStateCreateInfoEXT into it. The VkPipelineVertexInputDivisorStateCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructPipelineVertexInputDivisorStateCreateInfoEXT :: VkPipelineVertexInputDivisorStateCreateInfoEXT -> IO PipelineVertexInputDivisorStateCreateInfoEXT

A function to read a VkPipelineVertexInputDivisorStateCreateInfoEXT and all additional structures in the pointer chain into a PipelineVertexInputDivisorStateCreateInfoEXT.

data PipelineVertexInputDivisorStateCreateInfoEXT

VkPipelineVertexInputDivisorStateCreateInfoEXT - Structure specifying vertex attributes assignment during instanced rendering

Valid Usage (Implicit)

See Also

VkStructureType, VkVertexInputBindingDivisorDescriptionEXT

Constructors

PipelineVertexInputDivisorStateCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

vertexBindingDivisors :: Vector VertexInputBindingDivisorDescriptionEXT

withCStructVertexInputBindingDivisorDescriptionEXT :: VertexInputBindingDivisorDescriptionEXT -> (VkVertexInputBindingDivisorDescriptionEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkVertexInputBindingDivisorDescriptionEXT and marshal a VertexInputBindingDivisorDescriptionEXT into it. The VkVertexInputBindingDivisorDescriptionEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructVertexInputBindingDivisorDescriptionEXT :: VkVertexInputBindingDivisorDescriptionEXT -> IO VertexInputBindingDivisorDescriptionEXT

A function to read a VkVertexInputBindingDivisorDescriptionEXT and all additional structures in the pointer chain into a VertexInputBindingDivisorDescriptionEXT.

data VertexInputBindingDivisorDescriptionEXT

VkVertexInputBindingDivisorDescriptionEXT - Structure specifying a divisor used in instanced rendering

Description

If this structure is not used to define a divisor value for an attribute then the divisor has a logical default value of 1.

Valid Usage

binding must be less than VkPhysicalDeviceLimits::maxVertexInputBindings

If the vertexAttributeInstanceRateZeroDivisor feature is not enabled, divisor must not be 0

If the vertexAttributeInstanceRateDivisor feature is not enabled, divisor must be 1

divisor must be a value between 0 and VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT::maxVertexAttribDivisor, inclusive.

VkVertexInputBindingDescription::inputRate must be of type VK_VERTEX_INPUT_RATE_INSTANCE for this binding.

See Also

VkPipelineVertexInputDivisorStateCreateInfoEXT

Constructors

VertexInputBindingDivisorDescriptionEXT

Fields

binding :: Word32

divisor :: Word32

pattern EXT_VERTEX_ATTRIBUTE_DIVISOR_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_VERTEX_ATTRIBUTE_DIVISOR_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_VERTEX_ATTRIBUTE_DIVISOR_PROPERTIES_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_DIVISOR_STATE_CREATE_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_VERTEX_ATTRIBUTE_DIVISOR_FEATURES_EXT :: a ~ StructureType => a