Graphics.Vulkan.Extensions.VK_FUCHSIA_imagepipe_surface

Documentation

type ImagePipeSurfaceCreateFlagsFUCHSIA = VkImagePipeSurfaceCreateFlagsFUCHSIA

withCStructImagePipeSurfaceCreateInfoFUCHSIA :: ImagePipeSurfaceCreateInfoFUCHSIA -> (VkImagePipeSurfaceCreateInfoFUCHSIA -> IO a) -> IO a

A function to temporarily allocate memory for a VkImagePipeSurfaceCreateInfoFUCHSIA and marshal a ImagePipeSurfaceCreateInfoFUCHSIA into it. The VkImagePipeSurfaceCreateInfoFUCHSIA is only valid inside the provided computation and must not be returned out of it.

fromCStructImagePipeSurfaceCreateInfoFUCHSIA :: VkImagePipeSurfaceCreateInfoFUCHSIA -> IO ImagePipeSurfaceCreateInfoFUCHSIA

A function to read a VkImagePipeSurfaceCreateInfoFUCHSIA and all additional structures in the pointer chain into a ImagePipeSurfaceCreateInfoFUCHSIA.

data ImagePipeSurfaceCreateInfoFUCHSIA

VkImagePipeSurfaceCreateInfoFUCHSIA - Structure specifying parameters of a newly created ImagePipe surface object

Valid Usage (Implicit)

See Also

VkImagePipeSurfaceCreateFlagsFUCHSIA, VkStructureType, vkCreateImagePipeSurfaceFUCHSIA

Constructors

ImagePipeSurfaceCreateInfoFUCHSIA

Fields

next :: Maybe SomeVkStruct

flags :: ImagePipeSurfaceCreateFlagsFUCHSIA

imagePipeHandle :: Zx_handle_t

createImagePipeSurfaceFUCHSIA :: Instance -> ImagePipeSurfaceCreateInfoFUCHSIA -> Maybe AllocationCallbacks -> IO SurfaceKHR

vkCreateImagePipeSurfaceFUCHSIA - Create a VkSurfaceKHR object for a Fuchsia ImagePipe

Parameters

instance is the instance to associate with the surface.

pCreateInfo is a pointer to an instance of the VkImagePipeSurfaceCreateInfoFUCHSIA structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkImagePipeSurfaceCreateInfoFUCHSIA structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkImagePipeSurfaceCreateInfoFUCHSIA, VkInstance, VkSurfaceKHR

pattern FUCHSIA_IMAGEPIPE_SURFACE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern FUCHSIA_IMAGEPIPE_SURFACE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_IMAGEPIPE_SURFACE_CREATE_INFO_FUCHSIA :: a ~ StructureType => a