Graphics.Vulkan.Extensions.VK_GOOGLE_display_timing

Documentation

withCStructPastPresentationTimingGOOGLE :: PastPresentationTimingGOOGLE -> (VkPastPresentationTimingGOOGLE -> IO a) -> IO a

A function to temporarily allocate memory for a VkPastPresentationTimingGOOGLE and marshal a PastPresentationTimingGOOGLE into it. The VkPastPresentationTimingGOOGLE is only valid inside the provided computation and must not be returned out of it.

fromCStructPastPresentationTimingGOOGLE :: VkPastPresentationTimingGOOGLE -> IO PastPresentationTimingGOOGLE

A function to read a VkPastPresentationTimingGOOGLE and all additional structures in the pointer chain into a PastPresentationTimingGOOGLE.

data PastPresentationTimingGOOGLE

VkPastPresentationTimingGOOGLE - Structure containing timing information about a previously-presented image

Description

The results for a given swapchain and presentID are only returned once from vkGetPastPresentationTimingGOOGLE.

The application can use the VkPastPresentationTimingGOOGLE values to occasionally adjust its timing. For example, if actualPresentTime is later than expected (e.g. one refreshDuration late), the application may increase its target IPD to a higher multiple of refreshDuration (e.g. decrease its frame rate from 60Hz to 30Hz). If actualPresentTime and earliestPresentTime are consistently different, and if presentMargin is consistently large enough, the application may decrease its target IPD to a smaller multiple of refreshDuration (e.g. increase its frame rate from 30Hz to 60Hz). If actualPresentTime and earliestPresentTime are same, and if presentMargin is consistently high, the application may delay the start of its input-render-present loop in order to decrease the latency between user input and the corresponding present (always leaving some margin in case a new image takes longer to render than the previous image). An application that desires its target IPD to always be the same as refreshDuration, can also adjust features until actualPresentTime is never late and presentMargin is satisfactory.

See Also

vkGetPastPresentationTimingGOOGLE

Constructors

PastPresentationTimingGOOGLE

Fields

presentID :: Word32

desiredPresentTime :: Word64

actualPresentTime :: Word64

earliestPresentTime :: Word64

presentMargin :: Word64

withCStructPresentTimeGOOGLE :: PresentTimeGOOGLE -> (VkPresentTimeGOOGLE -> IO a) -> IO a

A function to temporarily allocate memory for a VkPresentTimeGOOGLE and marshal a PresentTimeGOOGLE into it. The VkPresentTimeGOOGLE is only valid inside the provided computation and must not be returned out of it.

fromCStructPresentTimeGOOGLE :: VkPresentTimeGOOGLE -> IO PresentTimeGOOGLE

A function to read a VkPresentTimeGOOGLE and all additional structures in the pointer chain into a PresentTimeGOOGLE.

data PresentTimeGOOGLE

VkPresentTimeGOOGLE - The earliest time image should be presented

See Also

VkPresentTimesInfoGOOGLE

Constructors

PresentTimeGOOGLE

Fields

presentID :: Word32

desiredPresentTime :: Word64

withCStructPresentTimesInfoGOOGLE :: PresentTimesInfoGOOGLE -> (VkPresentTimesInfoGOOGLE -> IO a) -> IO a

A function to temporarily allocate memory for a VkPresentTimesInfoGOOGLE and marshal a PresentTimesInfoGOOGLE into it. The VkPresentTimesInfoGOOGLE is only valid inside the provided computation and must not be returned out of it.

fromCStructPresentTimesInfoGOOGLE :: VkPresentTimesInfoGOOGLE -> IO PresentTimesInfoGOOGLE

A function to read a VkPresentTimesInfoGOOGLE and all additional structures in the pointer chain into a PresentTimesInfoGOOGLE.

data PresentTimesInfoGOOGLE

VkPresentTimesInfoGOOGLE - The earliest time each image should be presented

Valid Usage

swapchainCount must be the same value as VkPresentInfoKHR::swapchainCount, where VkPresentInfoKHR is in the pNext chain of this VkPresentTimesInfoGOOGLE structure.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PRESENT_TIMES_INFO_GOOGLE

If pTimes is not NULL, pTimes must be a valid pointer to an array of swapchainCount VkPresentTimeGOOGLE structures

swapchainCount must be greater than 0

See Also

VkPresentTimeGOOGLE, VkStructureType

Constructors

PresentTimesInfoGOOGLE

Fields

next :: Maybe SomeVkStruct

times :: Maybe (Vector PresentTimeGOOGLE)

withCStructRefreshCycleDurationGOOGLE :: RefreshCycleDurationGOOGLE -> (VkRefreshCycleDurationGOOGLE -> IO a) -> IO a

A function to temporarily allocate memory for a VkRefreshCycleDurationGOOGLE and marshal a RefreshCycleDurationGOOGLE into it. The VkRefreshCycleDurationGOOGLE is only valid inside the provided computation and must not be returned out of it.

fromCStructRefreshCycleDurationGOOGLE :: VkRefreshCycleDurationGOOGLE -> IO RefreshCycleDurationGOOGLE

A function to read a VkRefreshCycleDurationGOOGLE and all additional structures in the pointer chain into a RefreshCycleDurationGOOGLE.

data RefreshCycleDurationGOOGLE

VkRefreshCycleDurationGOOGLE - Structure containing the RC duration of a display

See Also

vkGetRefreshCycleDurationGOOGLE

Constructors

RefreshCycleDurationGOOGLE

Fields

refreshDuration :: Word64

getNumPastPresentationTimingGOOGLE :: Device -> SwapchainKHR -> IO (VkResult, Word32)

vkGetPastPresentationTimingGOOGLE - Obtain timing of a previously-presented image

Parameters

device is the device associated with swapchain.

swapchain is the swapchain to obtain presentation timing information duration for.

pPresentationTimingCount is a pointer to an integer related to the number of VkPastPresentationTimingGOOGLE structures to query, as described below.

pPresentationTimings is either NULL or a pointer to an array of VkPastPresentationTimingGOOGLE structures.

Description

If pPresentationTimings is NULL, then the number of newly-available timing records for the given swapchain is returned in pPresentationTimingCount. Otherwise, pPresentationTimingCount must point to a variable set by the user to the number of elements in the pPresentationTimings array, and on return the variable is overwritten with the number of structures actually written to pPresentationTimings. If the value of pPresentationTimingCount is less than the number of newly-available timing records, at most pPresentationTimingCount structures will be written. If pPresentationTimingCount is smaller than the number of newly-available timing records for the given swapchain, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

swapchain must be a valid VkSwapchainKHR handle

pPresentationTimingCount must be a valid pointer to a uint32_t value

If the value referenced by pPresentationTimingCount is not 0, and pPresentationTimings is not NULL, pPresentationTimings must be a valid pointer to an array of pPresentationTimingCount VkPastPresentationTimingGOOGLE structures

Both of device, and swapchain must have been created, allocated, or retrieved from the same VkInstance

Host Synchronization

Host access to swapchain must be externally synchronized

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_DEVICE_LOST

VK_ERROR_OUT_OF_DATE_KHR

VK_ERROR_SURFACE_LOST_KHR

See Also

VkDevice, VkPastPresentationTimingGOOGLE, VkSwapchainKHR

getPastPresentationTimingGOOGLE :: Device -> SwapchainKHR -> Word32 -> IO (VkResult, Vector PastPresentationTimingGOOGLE)

vkGetPastPresentationTimingGOOGLE - Obtain timing of a previously-presented image

Parameters

device is the device associated with swapchain.

swapchain is the swapchain to obtain presentation timing information duration for.

pPresentationTimingCount is a pointer to an integer related to the number of VkPastPresentationTimingGOOGLE structures to query, as described below.

pPresentationTimings is either NULL or a pointer to an array of VkPastPresentationTimingGOOGLE structures.

Description

If pPresentationTimings is NULL, then the number of newly-available timing records for the given swapchain is returned in pPresentationTimingCount. Otherwise, pPresentationTimingCount must point to a variable set by the user to the number of elements in the pPresentationTimings array, and on return the variable is overwritten with the number of structures actually written to pPresentationTimings. If the value of pPresentationTimingCount is less than the number of newly-available timing records, at most pPresentationTimingCount structures will be written. If pPresentationTimingCount is smaller than the number of newly-available timing records for the given swapchain, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

swapchain must be a valid VkSwapchainKHR handle

pPresentationTimingCount must be a valid pointer to a uint32_t value

If the value referenced by pPresentationTimingCount is not 0, and pPresentationTimings is not NULL, pPresentationTimings must be a valid pointer to an array of pPresentationTimingCount VkPastPresentationTimingGOOGLE structures

Both of device, and swapchain must have been created, allocated, or retrieved from the same VkInstance

Host Synchronization

Host access to swapchain must be externally synchronized

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_DEVICE_LOST

VK_ERROR_OUT_OF_DATE_KHR

VK_ERROR_SURFACE_LOST_KHR

See Also

VkDevice, VkPastPresentationTimingGOOGLE, VkSwapchainKHR

getAllPastPresentationTimingGOOGLE :: Device -> SwapchainKHR -> IO (Vector PastPresentationTimingGOOGLE)

Returns all the values available from getPastPresentationTimingGOOGLE.

getRefreshCycleDurationGOOGLE :: Device -> SwapchainKHR -> IO RefreshCycleDurationGOOGLE

vkGetRefreshCycleDurationGOOGLE - Obtain the RC duration of the PE’s display

Parameters

device is the device associated with swapchain.

swapchain is the swapchain to obtain the refresh duration for.

pDisplayTimingProperties is a pointer to an instance of the VkRefreshCycleDurationGOOGLE structure.

Valid Usage (Implicit)

device must be a valid VkDevice handle

swapchain must be a valid VkSwapchainKHR handle

pDisplayTimingProperties must be a valid pointer to a VkRefreshCycleDurationGOOGLE structure

Both of device, and swapchain must have been created, allocated, or retrieved from the same VkInstance

Host Synchronization

Host access to swapchain must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_DEVICE_LOST

VK_ERROR_SURFACE_LOST_KHR

See Also

VkDevice, VkRefreshCycleDurationGOOGLE, VkSwapchainKHR

pattern GOOGLE_DISPLAY_TIMING_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern GOOGLE_DISPLAY_TIMING_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PRESENT_TIMES_INFO_GOOGLE :: a ~ StructureType => a