Graphics.Vulkan.Extensions.VK_INTEL_performance_query

Documentation

initializePerformanceApiINTEL :: Device -> ("initializeInfo" ::: InitializePerformanceApiInfoINTEL) -> IO ()

vkInitializePerformanceApiINTEL - Initialize a device for performance queries

Parameters

Device is the logical device used for the queries.

pInitializeInfo is a pointer to a InitializePerformanceApiInfoINTEL structure specifying initialization parameters.

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

Device, InitializePerformanceApiInfoINTEL

uninitializePerformanceApiINTEL :: Device -> IO ()

vkUninitializePerformanceApiINTEL - Uninitialize a device for performance queries

Parameters

Device is the logical device used for the queries.

Valid Usage (Implicit)

See Also

Device

cmdSetPerformanceMarkerINTEL :: CommandBuffer -> PerformanceMarkerInfoINTEL -> IO ()

vkCmdSetPerformanceMarkerINTEL - Markers

Parameters

The last marker set onto a command buffer before the end of a query will be part of the query result.

Valid Usage (Implicit)

CommandBuffer must be a valid CommandBuffer handle

pMarkerInfo must be a valid pointer to a valid PerformanceMarkerInfoINTEL structure

CommandBuffer must be in the recording state

The CommandPool that CommandBuffer was allocated from must support graphics, compute, or transfer operations

Host Synchronization

Host access to the CommandPool that CommandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute Transfer

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

CommandBuffer, PerformanceMarkerInfoINTEL

cmdSetPerformanceStreamMarkerINTEL :: CommandBuffer -> PerformanceStreamMarkerInfoINTEL -> IO ()

vkCmdSetPerformanceStreamMarkerINTEL - Markers

Valid Usage (Implicit)

CommandBuffer must be a valid CommandBuffer handle

pMarkerInfo must be a valid pointer to a valid PerformanceStreamMarkerInfoINTEL structure

CommandBuffer must be in the recording state

The CommandPool that CommandBuffer was allocated from must support graphics, compute, or transfer operations

Host Synchronization

Host access to the CommandPool that CommandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute Transfer

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

CommandBuffer, PerformanceStreamMarkerInfoINTEL

cmdSetPerformanceOverrideINTEL :: CommandBuffer -> PerformanceOverrideInfoINTEL -> IO ()

vkCmdSetPerformanceOverrideINTEL - Performance override settings

Parameters

CommandBuffer is the command buffer where the override takes place.

pOverrideInfo is a pointer to a PerformanceOverrideInfoINTEL structure selecting the parameter to override.

Valid Usage

pOverrideInfo must not be used with a PerformanceOverrideTypeINTEL that is not reported available by getPerformanceParameterINTEL.

Valid Usage (Implicit)

CommandBuffer must be a valid CommandBuffer handle

pOverrideInfo must be a valid pointer to a valid PerformanceOverrideInfoINTEL structure

CommandBuffer must be in the recording state

The CommandPool that CommandBuffer was allocated from must support graphics, compute, or transfer operations

Host Synchronization

Host access to the CommandPool that CommandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute Transfer

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

CommandBuffer, PerformanceOverrideInfoINTEL

acquirePerformanceConfigurationINTEL :: Device -> PerformanceConfigurationAcquireInfoINTEL -> IO PerformanceConfigurationINTEL

vkAcquirePerformanceConfigurationINTEL - Acquire the performance query capability

Parameters

Device is the logical device that the performance query commands will be submitted to.

pAcquireInfo is a pointer to a PerformanceConfigurationAcquireInfoINTEL structure, specifying the performance configuration to acquire.

pConfiguration is a pointer to a PerformanceConfigurationINTEL handle in which the resulting configuration object is returned.

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

Device, PerformanceConfigurationAcquireInfoINTEL, PerformanceConfigurationINTEL

releasePerformanceConfigurationINTEL :: Device -> PerformanceConfigurationINTEL -> IO ()

vkReleasePerformanceConfigurationINTEL - Release a configuration to capture performance data

Parameters

Device is the device associated to the configuration object to release.

configuration is the configuration object to release.

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

Device, PerformanceConfigurationINTEL

queueSetPerformanceConfigurationINTEL :: Queue -> PerformanceConfigurationINTEL -> IO ()

vkQueueSetPerformanceConfigurationINTEL - Set a performance query

Parameters

Queue is the queue on which the configuration will be used.

configuration is the configuration to use.

Valid Usage (Implicit)

Queue must be a valid Queue handle

configuration must be a valid PerformanceConfigurationINTEL handle

Both of configuration, and Queue must have been created, allocated, or retrieved from the same Device

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

- 	 - 	 Any 	 -

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

PerformanceConfigurationINTEL, Queue

getPerformanceParameterINTEL :: Device -> PerformanceParameterTypeINTEL -> IO PerformanceValueINTEL

vkGetPerformanceParameterINTEL - Query performance capabilities of the device

Parameters

Device is the logical device to query.

parameter is the parameter to query.

pValue is a pointer to a PerformanceValueINTEL structure in which the type and value of the parameter are returned.

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

Device, PerformanceParameterTypeINTEL, PerformanceValueINTEL

data PerformanceValueINTEL

VkPerformanceValueINTEL - Container for value and types of parameters that can be queried

Valid Usage (Implicit)

See Also

PerformanceValueDataINTEL, PerformanceValueTypeINTEL, getPerformanceParameterINTEL

Constructors

PerformanceValueINTEL

Fields

type' :: PerformanceValueTypeINTEL

type must be a valid PerformanceValueTypeINTEL value

data' :: PerformanceValueDataINTEL

data must be a valid PerformanceValueDataINTEL union

data InitializePerformanceApiInfoINTEL

VkInitializePerformanceApiInfoINTEL - Structure specifying parameters of initialize of the device

Valid Usage (Implicit)

See Also

StructureType, initializePerformanceApiINTEL

Constructors

InitializePerformanceApiInfoINTEL

Fields

userData :: Ptr ()

pUserData is a pointer for application data.

data QueryPoolCreateInfoINTEL

VkQueryPoolCreateInfoINTEL - Structure specifying parameters to create a pool of performance queries

Members

To create a pool for Intel performance queries, set 'Graphics.Vulkan.Core10.Query.QueryPoolCreateInfo'::'Graphics.Vulkan.Core10.Enums.QueryType.QueryType' to QUERY_TYPE_PERFORMANCE_QUERY_INTEL and add a QueryPoolCreateInfoINTEL structure to the pNext chain of the QueryPoolCreateInfo structure.

Valid Usage (Implicit)

See Also

QueryPoolSamplingModeINTEL, StructureType

Constructors

QueryPoolCreateInfoINTEL

Fields

performanceCountersSampling :: QueryPoolSamplingModeINTEL

performanceCountersSampling must be a valid QueryPoolSamplingModeINTEL value

data PerformanceMarkerInfoINTEL

VkPerformanceMarkerInfoINTEL - Structure specifying performance markers

Valid Usage (Implicit)

See Also

StructureType, cmdSetPerformanceMarkerINTEL

Constructors

PerformanceMarkerInfoINTEL

Fields

marker :: Word64

marker is the marker value that will be recorded into the opaque query results.

data PerformanceStreamMarkerInfoINTEL

VkPerformanceStreamMarkerInfoINTEL - Structure specifying stream performance markers

Valid Usage

The value written by the application into marker must only used the valid bits as reported by getPerformanceParameterINTEL with the PERFORMANCE_PARAMETER_TYPE_STREAM_MARKER_VALID_BITS_INTEL.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PERFORMANCE_STREAM_MARKER_INFO_INTEL

pNext must be NULL

See Also

StructureType, cmdSetPerformanceStreamMarkerINTEL

Constructors

PerformanceStreamMarkerInfoINTEL

Fields

marker :: Word32

marker is the marker value that will be recorded into the reports consumed by an external application.

data PerformanceOverrideInfoINTEL

VkPerformanceOverrideInfoINTEL - Performance override info

Valid Usage (Implicit)

See Also

Bool32, PerformanceOverrideTypeINTEL, StructureType, cmdSetPerformanceOverrideINTEL

Constructors

PerformanceOverrideInfoINTEL

Fields

type' :: PerformanceOverrideTypeINTEL

type must be a valid PerformanceOverrideTypeINTEL value

enable :: Bool

enable defines whether the override is enabled.

parameter :: Word64

parameter is a potential required parameter for the override.

data PerformanceConfigurationAcquireInfoINTEL

VkPerformanceConfigurationAcquireInfoINTEL - Acquire a configuration to capture performance data

Valid Usage (Implicit)

See Also

PerformanceConfigurationTypeINTEL, StructureType, acquirePerformanceConfigurationINTEL

Constructors

PerformanceConfigurationAcquireInfoINTEL

Fields

type' :: PerformanceConfigurationTypeINTEL

type must be a valid PerformanceConfigurationTypeINTEL value

data PerformanceValueDataINTEL

Constructors

Value32 Word32

Value64 Word64

ValueFloat Float

ValueBool Bool

ValueString ByteString

peekPerformanceValueDataINTEL :: PerformanceValueTypeINTEL -> Ptr PerformanceValueDataINTEL -> IO PerformanceValueDataINTEL

newtype PerformanceConfigurationTypeINTEL

VkPerformanceConfigurationTypeINTEL - Type of performance configuration

See Also

PerformanceConfigurationAcquireInfoINTEL

Constructors

PerformanceConfigurationTypeINTEL Int32

Bundled Patterns

pattern PERFORMANCE_CONFIGURATION_TYPE_COMMAND_QUEUE_METRICS_DISCOVERY_ACTIVATED_INTEL :: PerformanceConfigurationTypeINTEL

newtype QueryPoolSamplingModeINTEL

VkQueryPoolSamplingModeINTEL - Enum specifying how performance queries should be captured

See Also

QueryPoolCreateInfoINTEL

Constructors

QueryPoolSamplingModeINTEL Int32

Bundled Patterns

pattern QUERY_POOL_SAMPLING_MODE_MANUAL_INTEL :: QueryPoolSamplingModeINTEL
QUERY_POOL_SAMPLING_MODE_MANUAL_INTEL is the default mode in which the application calls cmdBeginQuery and cmdEndQuery to record performance data.

newtype PerformanceOverrideTypeINTEL

VkPerformanceOverrideTypeINTEL - Performance override type

See Also

PerformanceOverrideInfoINTEL

Constructors

PerformanceOverrideTypeINTEL Int32

Bundled Patterns

pattern PERFORMANCE_OVERRIDE_TYPE_NULL_HARDWARE_INTEL :: PerformanceOverrideTypeINTEL
PERFORMANCE_OVERRIDE_TYPE_NULL_HARDWARE_INTEL turns all rendering operations into noop.

pattern PERFORMANCE_OVERRIDE_TYPE_FLUSH_GPU_CACHES_INTEL :: PerformanceOverrideTypeINTEL
PERFORMANCE_OVERRIDE_TYPE_FLUSH_GPU_CACHES_INTEL stalls the stream of commands until all previously emitted commands have completed and all caches been flushed and invalidated.

newtype PerformanceParameterTypeINTEL

VkPerformanceParameterTypeINTEL - Parameters that can be queried

See Also

getPerformanceParameterINTEL

Constructors

PerformanceParameterTypeINTEL Int32

Bundled Patterns

pattern PERFORMANCE_PARAMETER_TYPE_HW_COUNTERS_SUPPORTED_INTEL :: PerformanceParameterTypeINTEL
PERFORMANCE_PARAMETER_TYPE_HW_COUNTERS_SUPPORTED_INTEL has a boolean result which tells whether hardware counters can be captured.

pattern PERFORMANCE_PARAMETER_TYPE_STREAM_MARKER_VALID_BITS_INTEL :: PerformanceParameterTypeINTEL
PERFORMANCE_PARAMETER_TYPE_STREAM_MARKER_VALID_BITS_INTEL has a 32 bits integer result which tells how many bits can be written into the PerformanceValueINTEL value.

newtype PerformanceValueTypeINTEL

VkPerformanceValueTypeINTEL - Type of the parameters that can be queried

See Also

PerformanceValueINTEL

Constructors

PerformanceValueTypeINTEL Int32

Bundled Patterns

pattern PERFORMANCE_VALUE_TYPE_UINT32_INTEL :: PerformanceValueTypeINTEL

pattern PERFORMANCE_VALUE_TYPE_UINT64_INTEL :: PerformanceValueTypeINTEL

pattern PERFORMANCE_VALUE_TYPE_FLOAT_INTEL :: PerformanceValueTypeINTEL

pattern PERFORMANCE_VALUE_TYPE_BOOL_INTEL :: PerformanceValueTypeINTEL

pattern PERFORMANCE_VALUE_TYPE_STRING_INTEL :: PerformanceValueTypeINTEL

type INTEL_PERFORMANCE_QUERY_SPEC_VERSION = 1

pattern INTEL_PERFORMANCE_QUERY_SPEC_VERSION :: forall a. Integral a => a

type INTEL_PERFORMANCE_QUERY_EXTENSION_NAME = "VK_INTEL_performance_query"

pattern INTEL_PERFORMANCE_QUERY_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype PerformanceConfigurationINTEL

VkPerformanceConfigurationINTEL - Device configuration for performance queries

See Also

acquirePerformanceConfigurationINTEL, queueSetPerformanceConfigurationINTEL, releasePerformanceConfigurationINTEL

Constructors

PerformanceConfigurationINTEL Word64