Graphics.Vulkan.Extensions.VK_KHR_android_surface

Documentation

type AndroidSurfaceCreateFlagsKHR = VkAndroidSurfaceCreateFlagsKHR

withCStructAndroidSurfaceCreateInfoKHR :: AndroidSurfaceCreateInfoKHR -> (VkAndroidSurfaceCreateInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkAndroidSurfaceCreateInfoKHR and marshal a AndroidSurfaceCreateInfoKHR into it. The VkAndroidSurfaceCreateInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructAndroidSurfaceCreateInfoKHR :: VkAndroidSurfaceCreateInfoKHR -> IO AndroidSurfaceCreateInfoKHR

A function to read a VkAndroidSurfaceCreateInfoKHR and all additional structures in the pointer chain into a AndroidSurfaceCreateInfoKHR.

data AndroidSurfaceCreateInfoKHR

VkAndroidSurfaceCreateInfoKHR - Structure specifying parameters of a newly created Android surface object

Valid Usage (Implicit)

See Also

VkAndroidSurfaceCreateFlagsKHR, VkStructureType, vkCreateAndroidSurfaceKHR

Constructors

AndroidSurfaceCreateInfoKHR

Fields

next :: Maybe SomeVkStruct

flags :: AndroidSurfaceCreateFlagsKHR

window :: Ptr ANativeWindow

createAndroidSurfaceKHR :: Instance -> AndroidSurfaceCreateInfoKHR -> Maybe AllocationCallbacks -> IO SurfaceKHR

vkCreateAndroidSurfaceKHR - Create a VkSurfaceKHR object for an Android native window

Parameters

instance is the instance to associate the surface with.

pCreateInfo is a pointer to an instance of the VkAndroidSurfaceCreateInfoKHR structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface object is returned.

Description

During the lifetime of a surface created using a particular ANativeWindow handle any attempts to create another surface for the same ANativeWindow and any attempts to connect to the same ANativeWindow through other platform mechanisms will fail.

Note

In particular, only one VkSurfaceKHR can exist at a time for a given window. Similarly, a native window cannot be used by both a VkSurfaceKHR and EGLSurface simultaneously.

If successful, vkCreateAndroidSurfaceKHR increments the ANativeWindow’s reference count, and vkDestroySurfaceKHR will decrement it.

On Android, when a swapchain’s imageExtent does not match the surface’s currentExtent, the presentable images will be scaled to the surface’s dimensions during presentation. minImageExtent is (1,1), and maxImageExtent is the maximum image size supported by the consumer. For the system compositor, currentExtent is the window size (i.e. the consumer’s preferred size).

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkAndroidSurfaceCreateInfoKHR structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_NATIVE_WINDOW_IN_USE_KHR

See Also

VkAllocationCallbacks, VkAndroidSurfaceCreateInfoKHR, VkInstance, VkSurfaceKHR

pattern KHR_ANDROID_SURFACE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_ANDROID_SURFACE_SPEC_VERSION :: Integral a => a

data ANativeWindow

Opaque data

pattern STRUCTURE_TYPE_ANDROID_SURFACE_CREATE_INFO_KHR :: a ~ StructureType => a