Graphics.Vulkan.Extensions.VK_KHR_create_renderpass2

Documentation

withCStructAttachmentDescription2KHR :: AttachmentDescription2KHR -> (VkAttachmentDescription2KHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkAttachmentDescription2KHR and marshal a AttachmentDescription2KHR into it. The VkAttachmentDescription2KHR is only valid inside the provided computation and must not be returned out of it.

fromCStructAttachmentDescription2KHR :: VkAttachmentDescription2KHR -> IO AttachmentDescription2KHR

A function to read a VkAttachmentDescription2KHR and all additional structures in the pointer chain into a AttachmentDescription2KHR.

data AttachmentDescription2KHR

VkAttachmentDescription2KHR - Structure specifying an attachment description

Description

Parameters defined by this structure with the same name as those in VkAttachmentDescription have the identical effect to those parameters.

Valid Usage (Implicit)

See Also

VkAttachmentDescriptionFlags, VkAttachmentLoadOp, VkAttachmentStoreOp, VkFormat, VkImageLayout, VkRenderPassCreateInfo2KHR, VkSampleCountFlagBits, VkStructureType

Constructors

AttachmentDescription2KHR

Fields

next :: Maybe SomeVkStruct

flags :: AttachmentDescriptionFlags

format :: Format

samples :: SampleCountFlagBits

loadOp :: AttachmentLoadOp

storeOp :: AttachmentStoreOp

stencilLoadOp :: AttachmentLoadOp

stencilStoreOp :: AttachmentStoreOp

initialLayout :: ImageLayout

finalLayout :: ImageLayout

withCStructAttachmentReference2KHR :: AttachmentReference2KHR -> (VkAttachmentReference2KHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkAttachmentReference2KHR and marshal a AttachmentReference2KHR into it. The VkAttachmentReference2KHR is only valid inside the provided computation and must not be returned out of it.

fromCStructAttachmentReference2KHR :: VkAttachmentReference2KHR -> IO AttachmentReference2KHR

A function to read a VkAttachmentReference2KHR and all additional structures in the pointer chain into a AttachmentReference2KHR.

data AttachmentReference2KHR

VkAttachmentReference2KHR - Structure specifying an attachment reference

Description

Parameters defined by this structure with the same name as those in VkAttachmentReference have the identical effect to those parameters.

aspectMask has the same effect for the described attachment as VkInputAttachmentAspectReference::aspectMask has on each corresponding attachment. It is ignored when this structure is used to describe anything other than an input attachment reference.

Valid Usage

If attachment is not VK_ATTACHMENT_UNUSED, layout must not be VK_IMAGE_LAYOUT_UNDEFINED or VK_IMAGE_LAYOUT_PREINITIALIZED

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_ATTACHMENT_REFERENCE_2_KHR

layout must be a valid VkImageLayout value

See Also

VkImageAspectFlags, VkImageLayout, VkStructureType, VkSubpassDescription2KHR, VkSubpassDescriptionDepthStencilResolveKHR

Constructors

AttachmentReference2KHR

Fields

next :: Maybe SomeVkStruct

attachment :: Word32

layout :: ImageLayout

aspectMask :: ImageAspectFlags

withCStructRenderPassCreateInfo2KHR :: RenderPassCreateInfo2KHR -> (VkRenderPassCreateInfo2KHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkRenderPassCreateInfo2KHR and marshal a RenderPassCreateInfo2KHR into it. The VkRenderPassCreateInfo2KHR is only valid inside the provided computation and must not be returned out of it.

fromCStructRenderPassCreateInfo2KHR :: VkRenderPassCreateInfo2KHR -> IO RenderPassCreateInfo2KHR

A function to read a VkRenderPassCreateInfo2KHR and all additional structures in the pointer chain into a RenderPassCreateInfo2KHR.

data RenderPassCreateInfo2KHR

VkRenderPassCreateInfo2KHR - Structure specifying parameters of a newly created render pass

Description

Parameters defined by this structure with the same name as those in VkRenderPassCreateInfo have the identical effect to those parameters; the child structures are variants of those used in VkRenderPassCreateInfo which include sType and pNext parameters, allowing them to be extended.

If the VkSubpassDescription2KHR::viewMask member of any element of pSubpasses is not zero, multiview functionality is considered to be enabled for this render pass.

correlatedViewMaskCount and pCorrelatedViewMasks have the same effect as VkRenderPassMultiviewCreateInfo::correlationMaskCount and VkRenderPassMultiviewCreateInfo::pCorrelationMasks, respectively.

Valid Usage

If any two subpasses operate on attachments with overlapping ranges of the same VkDeviceMemory object, and at least one subpass writes to that area of VkDeviceMemory, a subpass dependency must be included (either directly or via some intermediate subpasses) between them

If the attachment member of any element of pInputAttachments, pColorAttachments, pResolveAttachments or pDepthStencilAttachment, or the attachment indexed by any element of pPreserveAttachments in any given element of pSubpasses is bound to a range of a VkDeviceMemory object that overlaps with any other attachment in any subpass (including the same subpass), the VkAttachmentDescription2KHR structures describing them must include VK_ATTACHMENT_DESCRIPTION_MAY_ALIAS_BIT in flags

If the attachment member of any element of pInputAttachments, pColorAttachments, pResolveAttachments or pDepthStencilAttachment, or any element of pPreserveAttachments in any given element of pSubpasses is not VK_ATTACHMENT_UNUSED, it must be less than attachmentCount

For any member of pAttachments with a loadOp equal to VK_ATTACHMENT_LOAD_OP_CLEAR, the first use of that attachment must not specify a layout equal to VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL, VK_IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL, or VK_IMAGE_LAYOUT_DEPTH_READ_ONLY_STENCIL_ATTACHMENT_OPTIMAL

For any member of pAttachments with a stencilLoadOp equal to VK_ATTACHMENT_LOAD_OP_CLEAR, the first use of that attachment must not specify a layout equal to VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL, VK_IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL, or VK_IMAGE_LAYOUT_DEPTH_ATTACHMENT_STENCIL_READ_ONLY_OPTIMAL.

For any element of pDependencies, if the srcSubpass is not VK_SUBPASS_EXTERNAL, all stage flags included in the srcStageMask member of that dependency must be a pipeline stage supported by the pipeline identified by the pipelineBindPoint member of the source subpass.

For any element of pDependencies, if the dstSubpass is not VK_SUBPASS_EXTERNAL, all stage flags included in the dstStageMask member of that dependency must be a pipeline stage supported by the pipeline identified by the pipelineBindPoint member of the source subpass.

The set of bits included in any element of pCorrelatedViewMasks must not overlap with the set of bits included in any other element of pCorrelatedViewMasks

If the VkSubpassDescription2KHR::viewMask member of all elements of pSubpasses is 0, correlatedViewMaskCount must be 0

The VkSubpassDescription2KHR::viewMask member of all elements of pSubpasses must either all be 0, or all not be 0

If the VkSubpassDescription2KHR::viewMask member of all elements of pSubpasses is 0, the dependencyFlags member of any element of pDependencies must not include VK_DEPENDENCY_VIEW_LOCAL_BIT

For any element of pDependencies where its srcSubpass member equals its dstSubpass member, if the viewMask member of the corresponding element of pSubpasses includes more than one bit, its dependencyFlags member must include VK_DEPENDENCY_VIEW_LOCAL_BIT

The viewMask member must not include a bit at a position greater than the value of VkPhysicalDeviceLimits::maxFramebufferLayers

If the attachment member of any element of the pInputAttachments member of any element of pSubpasses is not VK_ATTACHMENT_UNUSED, the aspectMask member of that element of pInputAttachments must only include aspects that are present in images of the format specified by the element of pAttachments specified by attachment

The srcSubpass member of each element of pDependencies must be less than subpassCount

The dstSubpass member of each element of pDependencies must be less than subpassCount

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO_2_KHR

pNext must be NULL

If attachmentCount is not 0, pAttachments must be a valid pointer to an array of attachmentCount valid VkAttachmentDescription2KHR structures

pSubpasses must be a valid pointer to an array of subpassCount valid VkSubpassDescription2KHR structures

If dependencyCount is not 0, pDependencies must be a valid pointer to an array of dependencyCount valid VkSubpassDependency2KHR structures

If correlatedViewMaskCount is not 0, pCorrelatedViewMasks must be a valid pointer to an array of correlatedViewMaskCount uint32_t values

subpassCount must be greater than 0

See Also

VkAttachmentDescription2KHR, VkRenderPassCreateFlags, VkStructureType, VkSubpassDependency2KHR, VkSubpassDescription2KHR, vkCreateRenderPass2KHR

Constructors

RenderPassCreateInfo2KHR

Fields

next :: Maybe SomeVkStruct

flags :: RenderPassCreateFlags

attachments :: Vector AttachmentDescription2KHR

subpasses :: Vector SubpassDescription2KHR

dependencies :: Vector SubpassDependency2KHR

correlatedViewMasks :: Vector Word32

withCStructSubpassBeginInfoKHR :: SubpassBeginInfoKHR -> (VkSubpassBeginInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkSubpassBeginInfoKHR and marshal a SubpassBeginInfoKHR into it. The VkSubpassBeginInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructSubpassBeginInfoKHR :: VkSubpassBeginInfoKHR -> IO SubpassBeginInfoKHR

A function to read a VkSubpassBeginInfoKHR and all additional structures in the pointer chain into a SubpassBeginInfoKHR.

data SubpassBeginInfoKHR

VkSubpassBeginInfoKHR - Structure specifying subpass begin info

Valid Usage (Implicit)

See Also

VkStructureType, VkSubpassContents, vkCmdBeginRenderPass2KHR, vkCmdNextSubpass2KHR

Constructors

SubpassBeginInfoKHR

Fields

next :: Maybe SomeVkStruct

contents :: SubpassContents

withCStructSubpassDependency2KHR :: SubpassDependency2KHR -> (VkSubpassDependency2KHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkSubpassDependency2KHR and marshal a SubpassDependency2KHR into it. The VkSubpassDependency2KHR is only valid inside the provided computation and must not be returned out of it.

fromCStructSubpassDependency2KHR :: VkSubpassDependency2KHR -> IO SubpassDependency2KHR

A function to read a VkSubpassDependency2KHR and all additional structures in the pointer chain into a SubpassDependency2KHR.

data SubpassDependency2KHR

VkSubpassDependency2KHR - Structure specifying a subpass dependency

Description

Parameters defined by this structure with the same name as those in VkSubpassDependency have the identical effect to those parameters.

viewOffset has the same effect for the described subpass dependency as VkRenderPassMultiviewCreateInfo::pViewOffsets has on each corresponding subpass dependency.

Valid Usage

If srcSubpass is not VK_SUBPASS_EXTERNAL, srcStageMask must not include VK_PIPELINE_STAGE_HOST_BIT

If dstSubpass is not VK_SUBPASS_EXTERNAL, dstStageMask must not include VK_PIPELINE_STAGE_HOST_BIT

If the geometry shaders feature is not enabled, srcStageMask must not contain VK_PIPELINE_STAGE_GEOMETRY_SHADER_BIT

If the geometry shaders feature is not enabled, dstStageMask must not contain VK_PIPELINE_STAGE_GEOMETRY_SHADER_BIT

If the tessellation shaders feature is not enabled, srcStageMask must not contain VK_PIPELINE_STAGE_TESSELLATION_CONTROL_SHADER_BIT or VK_PIPELINE_STAGE_TESSELLATION_EVALUATION_SHADER_BIT

If the tessellation shaders feature is not enabled, dstStageMask must not contain VK_PIPELINE_STAGE_TESSELLATION_CONTROL_SHADER_BIT or VK_PIPELINE_STAGE_TESSELLATION_EVALUATION_SHADER_BIT

srcSubpass must be less than or equal to dstSubpass, unless one of them is VK_SUBPASS_EXTERNAL, to avoid cyclic dependencies and ensure a valid execution order

srcSubpass and dstSubpass must not both be equal to VK_SUBPASS_EXTERNAL

If srcSubpass is equal to dstSubpass, srcStageMask and dstStageMask must not set any bits that are neither VK_PIPELINE_STAGE_ALL_GRAPHICS_BIT, nor one of the graphics pipeline stages

If srcSubpass is equal to dstSubpass and not all of the stages in srcStageMask and dstStageMask are framebuffer-space stages, the logically latest pipeline stage in srcStageMask must be logically earlier than or equal to the logically earliest pipeline stage in dstStageMask

Any access flag included in srcAccessMask must be supported by one of the pipeline stages in srcStageMask, as specified in the table of supported access types

Any access flag included in dstAccessMask must be supported by one of the pipeline stages in dstStageMask, as specified in the table of supported access types

If dependencyFlags includes VK_DEPENDENCY_VIEW_LOCAL_BIT, srcSubpass must not be equal to VK_SUBPASS_EXTERNAL

If dependencyFlags includes VK_DEPENDENCY_VIEW_LOCAL_BIT, dstSubpass must not be equal to VK_SUBPASS_EXTERNAL

If srcSubpass equals dstSubpass, and srcStageMask and dstStageMask both include a framebuffer-space stage, then dependencyFlags must include VK_DEPENDENCY_BY_REGION_BIT

If viewOffset is not equal to 0, srcSubpass must not be equal to dstSubpass

If dependencyFlags does not include VK_DEPENDENCY_VIEW_LOCAL_BIT, viewOffset must be 0

If viewOffset is not 0, srcSubpass must not be equal to dstSubpass.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_SUBPASS_DEPENDENCY_2_KHR

srcStageMask must be a valid combination of VkPipelineStageFlagBits values

srcStageMask must not be 0

dstStageMask must be a valid combination of VkPipelineStageFlagBits values

dstStageMask must not be 0

srcAccessMask must be a valid combination of VkAccessFlagBits values

dstAccessMask must be a valid combination of VkAccessFlagBits values

dependencyFlags must be a valid combination of VkDependencyFlagBits values

See Also

VkAccessFlags, VkDependencyFlags, VkPipelineStageFlags, VkRenderPassCreateInfo2KHR, VkStructureType

Constructors

SubpassDependency2KHR

Fields

next :: Maybe SomeVkStruct

srcSubpass :: Word32

dstSubpass :: Word32

srcStageMask :: PipelineStageFlags

dstStageMask :: PipelineStageFlags

srcAccessMask :: AccessFlags

dstAccessMask :: AccessFlags

dependencyFlags :: DependencyFlags

viewOffset :: Int32

withCStructSubpassDescription2KHR :: SubpassDescription2KHR -> (VkSubpassDescription2KHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkSubpassDescription2KHR and marshal a SubpassDescription2KHR into it. The VkSubpassDescription2KHR is only valid inside the provided computation and must not be returned out of it.

fromCStructSubpassDescription2KHR :: VkSubpassDescription2KHR -> IO SubpassDescription2KHR

A function to read a VkSubpassDescription2KHR and all additional structures in the pointer chain into a SubpassDescription2KHR.

data SubpassDescription2KHR

VkSubpassDescription2KHR - Structure specifying a subpass description

Description

Parameters defined by this structure with the same name as those in VkSubpassDescription have the identical effect to those parameters.

viewMask has the same effect for the described subpass as VkRenderPassMultiviewCreateInfo::pViewMasks has on each corresponding subpass.

Valid Usage

pipelineBindPoint must be VK_PIPELINE_BIND_POINT_GRAPHICS

colorAttachmentCount must be less than or equal to VkPhysicalDeviceLimits::maxColorAttachments

If the first use of an attachment in this render pass is as an input attachment, and the attachment is not also used as a color or depth/stencil attachment in the same subpass, then loadOp must not be VK_ATTACHMENT_LOAD_OP_CLEAR

If pResolveAttachments is not NULL, for each resolve attachment that does not have the value VK_ATTACHMENT_UNUSED, the corresponding color attachment must not have the value VK_ATTACHMENT_UNUSED

If pResolveAttachments is not NULL, for each resolve attachment that is not VK_ATTACHMENT_UNUSED, the corresponding color attachment must not have a sample count of VK_SAMPLE_COUNT_1_BIT

If pResolveAttachments is not NULL, each resolve attachment that is not VK_ATTACHMENT_UNUSED must have a sample count of VK_SAMPLE_COUNT_1_BIT

Any given element of pResolveAttachments must have the same VkFormat as its corresponding color attachment

All attachments in pColorAttachments that are not VK_ATTACHMENT_UNUSED must have the same sample count

If neither the VK_AMD_mixed_attachment_samples nor the VK_NV_framebuffer_mixed_samples extensions are enabled, and if pDepthStencilAttachment is not VK_ATTACHMENT_UNUSED and any attachments in pColorAttachments are not VK_ATTACHMENT_UNUSED, they must have the same sample count

The attachment member of any element of pPreserveAttachments must not be VK_ATTACHMENT_UNUSED

Any given element of pPreserveAttachments must not also be an element of any other member of the subpass description

If any attachment is used by more than one VkAttachmentReference member, then each use must use the same layout

The aspectMask member of any element of pInputAttachments must be a valid combination of VkImageAspectFlagBits

The aspectMask member of any element of pInputAttachments must not be 0

The aspectMask member of each element of pInputAttachments must not include VK_IMAGE_ASPECT_METADATA_BIT

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_SUBPASS_DESCRIPTION_2_KHR

flags must be a valid combination of VkSubpassDescriptionFlagBits values

pipelineBindPoint must be a valid VkPipelineBindPoint value

If inputAttachmentCount is not 0, pInputAttachments must be a valid pointer to an array of inputAttachmentCount valid VkAttachmentReference2KHR structures

If colorAttachmentCount is not 0, pColorAttachments must be a valid pointer to an array of colorAttachmentCount valid VkAttachmentReference2KHR structures

If colorAttachmentCount is not 0, and pResolveAttachments is not NULL, pResolveAttachments must be a valid pointer to an array of colorAttachmentCount valid VkAttachmentReference2KHR structures

If pDepthStencilAttachment is not NULL, pDepthStencilAttachment must be a valid pointer to a valid VkAttachmentReference2KHR structure

If preserveAttachmentCount is not 0, pPreserveAttachments must be a valid pointer to an array of preserveAttachmentCount uint32_t values

See Also

VkAttachmentReference2KHR, VkPipelineBindPoint, VkRenderPassCreateInfo2KHR, VkStructureType, VkSubpassDescriptionFlags

Constructors

SubpassDescription2KHR

Fields

next :: Maybe SomeVkStruct

flags :: SubpassDescriptionFlags

pipelineBindPoint :: PipelineBindPoint

viewMask :: Word32

inputAttachments :: Vector AttachmentReference2KHR

colorAttachments :: Vector AttachmentReference2KHR

resolveAttachments :: Maybe (Vector AttachmentReference2KHR)

depthStencilAttachment :: Maybe AttachmentReference2KHR

preserveAttachments :: Vector Word32

withCStructSubpassEndInfoKHR :: SubpassEndInfoKHR -> (VkSubpassEndInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkSubpassEndInfoKHR and marshal a SubpassEndInfoKHR into it. The VkSubpassEndInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructSubpassEndInfoKHR :: VkSubpassEndInfoKHR -> IO SubpassEndInfoKHR

A function to read a VkSubpassEndInfoKHR and all additional structures in the pointer chain into a SubpassEndInfoKHR.

data SubpassEndInfoKHR

VkSubpassEndInfoKHR - Structure specifying subpass end info

Valid Usage (Implicit)

See Also

VkStructureType, vkCmdEndRenderPass2KHR, vkCmdNextSubpass2KHR

Constructors

SubpassEndInfoKHR

Fields

next :: Maybe SomeVkStruct

cmdBeginRenderPass2KHR :: CommandBuffer -> RenderPassBeginInfo -> SubpassBeginInfoKHR -> IO ()

vkCmdBeginRenderPass2KHR - Begin a new render pass

Parameters

commandBuffer is the command buffer in which to record the command.

pRenderPassBegin is a pointer to a VkRenderPassBeginInfo structure (defined below) which indicates the render pass to begin an instance of, and the framebuffer the instance uses.

pSubpassBeginInfo is a pointer to a VkSubpassBeginInfoKHR structure which contains information about the subpass which is about to begin rendering.

Description

After beginning a render pass instance, the command buffer is ready to record the commands for the first subpass of that render pass.

Valid Usage

If any of the initialLayout or finalLayout member of the VkAttachmentDescription structures or the layout member of the VkAttachmentReference structures specified when creating the render pass specified in the renderPass member of pRenderPassBegin is VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL then the corresponding attachment image view of the framebuffer specified in the framebuffer member of pRenderPassBegin must have been created with a usage value including VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT

If any of the initialLayout or finalLayout member of the VkAttachmentDescription structures or the layout member of the VkAttachmentReference structures specified when creating the render pass specified in the renderPass member of pRenderPassBegin is VK_IMAGE_LAYOUT_DEPTH_READ_ONLY_STENCIL_ATTACHMENT_OPTIMAL, VK_IMAGE_LAYOUT_DEPTH_ATTACHMENT_STENCIL_READ_ONLY_OPTIMAL, VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL, or VK_IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL then the corresponding attachment image view of the framebuffer specified in the framebuffer member of pRenderPassBegin must have been created with a usage value including VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT

If any of the initialLayout or finalLayout member of the VkAttachmentDescription structures or the layout member of the VkAttachmentReference structures specified when creating the render pass specified in the renderPass member of pRenderPassBegin is VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL then the corresponding attachment image view of the framebuffer specified in the framebuffer member of pRenderPassBegin must have been created with a usage value including VK_IMAGE_USAGE_SAMPLED_BIT or VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT

If any of the initialLayout or finalLayout member of the VkAttachmentDescription structures or the layout member of the VkAttachmentReference structures specified when creating the render pass specified in the renderPass member of pRenderPassBegin is VK_IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL then the corresponding attachment image view of the framebuffer specified in the framebuffer member of pRenderPassBegin must have been created with a usage value including VK_IMAGE_USAGE_TRANSFER_SRC_BIT

If any of the initialLayout or finalLayout member of the VkAttachmentDescription structures or the layout member of the VkAttachmentReference structures specified when creating the render pass specified in the renderPass member of pRenderPassBegin is VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL then the corresponding attachment image view of the framebuffer specified in the framebuffer member of pRenderPassBegin must have been created with a usage value including VK_IMAGE_USAGE_TRANSFER_DST_BIT

If any of the initialLayout members of the VkAttachmentDescription structures specified when creating the render pass specified in the renderPass member of pRenderPassBegin is not VK_IMAGE_LAYOUT_UNDEFINED, then each such initialLayout must be equal to the current layout of the corresponding attachment image subresource of the framebuffer specified in the framebuffer member of pRenderPassBegin

The srcStageMask and dstStageMask members of any element of the pDependencies member of VkRenderPassCreateInfo used to create renderPass must be supported by the capabilities of the queue family identified by the queueFamilyIndex member of the VkCommandPoolCreateInfo used to create the command pool which commandBuffer was allocated from

For any attachment in framebuffer that is used by renderPass and is bound to memory locations that are also bound to another attachment used by renderPass, and if at least one of those uses causes either attachment to be written to, both attachments must have had the VK_ATTACHMENT_DESCRIPTION_MAY_ALIAS_BIT set

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pRenderPassBegin must be a valid pointer to a valid VkRenderPassBeginInfo structure

pSubpassBeginInfo must be a valid pointer to a valid VkSubpassBeginInfoKHR structure

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics operations

This command must only be called outside of a render pass instance

commandBuffer must be a primary VkCommandBuffer

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Outside         | Graphics        | Graphics        |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkRenderPassBeginInfo, VkSubpassBeginInfoKHR

cmdEndRenderPass2KHR :: CommandBuffer -> SubpassEndInfoKHR -> IO ()

vkCmdEndRenderPass2KHR - End the current render pass

Parameters

commandBuffer is the command buffer in which to end the current render pass instance.

pSubpassEndInfo is a pointer to a VkSubpassEndInfoKHR structure which contains information about how the previous subpass will be ended.

Description

vkCmdEndRenderPass2KHR is semantically identical to vkCmdEndRenderPass, except that it is extensible.

Valid Usage

The current subpass index must be equal to the number of subpasses in the render pass minus one

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pSubpassEndInfo must be a valid pointer to a valid VkSubpassEndInfoKHR structure

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics operations

This command must only be called inside of a render pass instance

commandBuffer must be a primary VkCommandBuffer

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Inside          | Graphics        | Graphics        |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkSubpassEndInfoKHR

cmdNextSubpass2KHR :: CommandBuffer -> SubpassBeginInfoKHR -> SubpassEndInfoKHR -> IO ()

vkCmdNextSubpass2KHR - Transition to the next subpass of a render pass

Parameters

commandBuffer is the command buffer in which to record the command.

pSubpassBeginInfo is a pointer to a VkSubpassBeginInfoKHR structure which contains information about the subpass which is about to begin rendering.

pSubpassEndInfo is a pointer to a VkSubpassEndInfoKHR structure which contains information about how the previous subpass will be ended.

Description

vkCmdNextSubpass2KHR is semantically identical to vkCmdNextSubpass, except that it is extensible, and that contents is provided as part of an extensible structure instead of as a flat parameter.

Valid Usage

The current subpass index must be less than the number of subpasses in the render pass minus one

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pSubpassBeginInfo must be a valid pointer to a valid VkSubpassBeginInfoKHR structure

pSubpassEndInfo must be a valid pointer to a valid VkSubpassEndInfoKHR structure

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics operations

This command must only be called inside of a render pass instance

commandBuffer must be a primary VkCommandBuffer

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Inside          | Graphics        | Graphics        |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkSubpassBeginInfoKHR, VkSubpassEndInfoKHR

createRenderPass2KHR :: Device -> RenderPassCreateInfo2KHR -> Maybe AllocationCallbacks -> IO RenderPass

vkCreateRenderPass2KHR - Create a new render pass object

Parameters

device is the logical device that creates the render pass.

pCreateInfo is a pointer to an instance of the VkRenderPassCreateInfo2KHR structure that describes the parameters of the render pass.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pRenderPass points to a VkRenderPass handle in which the resulting render pass object is returned.

Description

This command is functionally identical to vkCreateRenderPass, but includes extensible sub-structures that include sType and pNext parameters, allowing them to be more easily extended.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkRenderPassCreateInfo2KHR structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pRenderPass must be a valid pointer to a VkRenderPass handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDevice, VkRenderPass, VkRenderPassCreateInfo2KHR

pattern KHR_CREATE_RENDERPASS_2_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_CREATE_RENDERPASS_2_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_ATTACHMENT_DESCRIPTION_2_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_ATTACHMENT_REFERENCE_2_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SUBPASS_DESCRIPTION_2_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SUBPASS_DEPENDENCY_2_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO_2_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SUBPASS_BEGIN_INFO_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SUBPASS_END_INFO_KHR :: a ~ StructureType => a