Graphics.Vulkan.Extensions.VK_KHR_depth_stencil_resolve

Documentation

withCStructPhysicalDeviceDepthStencilResolvePropertiesKHR :: PhysicalDeviceDepthStencilResolvePropertiesKHR -> (VkPhysicalDeviceDepthStencilResolvePropertiesKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceDepthStencilResolvePropertiesKHR and marshal a PhysicalDeviceDepthStencilResolvePropertiesKHR into it. The VkPhysicalDeviceDepthStencilResolvePropertiesKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceDepthStencilResolvePropertiesKHR :: VkPhysicalDeviceDepthStencilResolvePropertiesKHR -> IO PhysicalDeviceDepthStencilResolvePropertiesKHR

A function to read a VkPhysicalDeviceDepthStencilResolvePropertiesKHR and all additional structures in the pointer chain into a PhysicalDeviceDepthStencilResolvePropertiesKHR.

data PhysicalDeviceDepthStencilResolvePropertiesKHR

VkPhysicalDeviceDepthStencilResolvePropertiesKHR - Structure describing depth/stencil resolve properties that can be supported by an implementation

Members

The members of the VkPhysicalDeviceDepthStencilResolvePropertiesKHR structure describe the following implementation-dependent limits:

Valid Usage (Implicit)

See Also

VkBool32, VkResolveModeFlagsKHR, VkStructureType

Constructors

PhysicalDeviceDepthStencilResolvePropertiesKHR

Fields

next :: Maybe SomeVkStruct

supportedDepthResolveModes :: ResolveModeFlagsKHR

supportedStencilResolveModes :: ResolveModeFlagsKHR

independentResolveNone :: Bool

independentResolve :: Bool

type ResolveModeFlagBitsKHR = VkResolveModeFlagBitsKHR

VkResolveModeFlagBitsKHR - Bitmask indicating supported depth and stencil resolve modes

See Also

VkResolveModeFlagsKHR, VkSubpassDescriptionDepthStencilResolveKHR

pattern RESOLVE_MODE_SAMPLE_ZERO_BIT_KHR :: a ~ ResolveModeFlagBitsKHR => a

VK_RESOLVE_MODE_SAMPLE_ZERO_BIT_KHR indicates that result of the resolve operation is equal to the value of sample 0.

pattern RESOLVE_MODE_AVERAGE_BIT_KHR :: a ~ ResolveModeFlagBitsKHR => a

VK_RESOLVE_MODE_AVERAGE_BIT_KHR indicates that result of the resolve operation is the average of the sample values.

pattern RESOLVE_MODE_MIN_BIT_KHR :: a ~ ResolveModeFlagBitsKHR => a

VK_RESOLVE_MODE_MIN_BIT_KHR indicates that result of the resolve operation is the minimum of the sample values.

pattern RESOLVE_MODE_MAX_BIT_KHR :: a ~ ResolveModeFlagBitsKHR => a

VK_RESOLVE_MODE_MAX_BIT_KHR indicates that result of the resolve operation is the maximum of the sample values.

pattern RESOLVE_MODE_NONE_KHR :: a ~ ResolveModeFlagBitsKHR => a

VK_RESOLVE_MODE_NONE_KHR indicates that no resolve operation is done.

type ResolveModeFlagsKHR = ResolveModeFlagBitsKHR

VkResolveModeFlagsKHR - Bitmask of VkResolveModeFlagBitsKHR

Description

VkResolveModeFlagsKHR is a bitmask type for setting a mask of zero or more VkResolveModeFlagBitsKHR.

See Also

VkPhysicalDeviceDepthStencilResolvePropertiesKHR, VkResolveModeFlagBitsKHR

withCStructSubpassDescriptionDepthStencilResolveKHR :: SubpassDescriptionDepthStencilResolveKHR -> (VkSubpassDescriptionDepthStencilResolveKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkSubpassDescriptionDepthStencilResolveKHR and marshal a SubpassDescriptionDepthStencilResolveKHR into it. The VkSubpassDescriptionDepthStencilResolveKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructSubpassDescriptionDepthStencilResolveKHR :: VkSubpassDescriptionDepthStencilResolveKHR -> IO SubpassDescriptionDepthStencilResolveKHR

A function to read a VkSubpassDescriptionDepthStencilResolveKHR and all additional structures in the pointer chain into a SubpassDescriptionDepthStencilResolveKHR.

data SubpassDescriptionDepthStencilResolveKHR

VkSubpassDescriptionDepthStencilResolveKHR - Structure specifying depth/stencil resolve operations for a subpass

Valid Usage

If pDepthStencilResolveAttachment is not NULL and does not have the value VK_ATTACHMENT_UNUSED, pDepthStencilAttachment must not have the value VK_ATTACHMENT_UNUSED

If pDepthStencilResolveAttachment is not NULL and does not have the value VK_ATTACHMENT_UNUSED, depthResolveMode and stencilResolveMode must not both be VK_RESOLVE_MODE_NONE_KHR

If pDepthStencilResolveAttachment is not NULL and does not have the value VK_ATTACHMENT_UNUSED, pDepthStencilAttachment must not have a sample count of VK_SAMPLE_COUNT_1_BIT

If pDepthStencilResolveAttachment is not NULL and does not have the value VK_ATTACHMENT_UNUSED, pDepthStencilResolveAttachment must have a sample count of VK_SAMPLE_COUNT_1_BIT

If pDepthStencilResolveAttachment is not NULL and does not have the value VK_ATTACHMENT_UNUSED then it must have a format whose features contain VK_FORMAT_FEATURE_DEPTH_STENCIL_ATTACHMENT_BIT

If the VkFormat of pDepthStencilResolveAttachment has a depth component, then the VkFormat of pDepthStencilAttachment must have a depth component with the same number of bits and numerical type

If the VkFormat of pDepthStencilResolveAttachment has a stencil component, then the VkFormat of pDepthStencilAttachment must have a stencil component with the same number of bits and numerical type

The value of depthResolveMode must be one of the bits set in VkPhysicalDeviceDepthStencilResolvePropertiesKHR::supportedDepthResolveModes or VK_RESOLVE_MODE_NONE_KHR

The value of stencilResolveMode must be one of the bits set in VkPhysicalDeviceDepthStencilResolvePropertiesKHR::supportedStencilResolveModes or VK_RESOLVE_MODE_NONE_KHR

If the VkFormat of pDepthStencilResolveAttachment has both depth and stencil components, VkPhysicalDeviceDepthStencilResolvePropertiesKHR::independentResolve is VK_FALSE, and VkPhysicalDeviceDepthStencilResolvePropertiesKHR::independentResolveNone is VK_FALSE, then the values of depthResolveMode and stencilResolveMode must be identical

If the VkFormat of pDepthStencilResolveAttachment has both depth and stencil components, VkPhysicalDeviceDepthStencilResolvePropertiesKHR::independentResolve is VK_FALSE and VkPhysicalDeviceDepthStencilResolvePropertiesKHR::independentResolveNone is VK_TRUE, then the values of depthResolveMode and stencilResolveMode must be identical or one of them must be VK_RESOLVE_MODE_NONE_KHR

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_SUBPASS_DESCRIPTION_DEPTH_STENCIL_RESOLVE_KHR

depthResolveMode must be a valid VkResolveModeFlagBitsKHR value

stencilResolveMode must be a valid VkResolveModeFlagBitsKHR value

If pDepthStencilResolveAttachment is not NULL, pDepthStencilResolveAttachment must be a valid pointer to a valid VkAttachmentReference2KHR structure

See Also

VkAttachmentReference2KHR, VkResolveModeFlagBitsKHR, VkStructureType

Constructors

SubpassDescriptionDepthStencilResolveKHR

Fields

next :: Maybe SomeVkStruct

depthResolveMode :: ResolveModeFlagBitsKHR

stencilResolveMode :: ResolveModeFlagBitsKHR

depthStencilResolveAttachment :: Maybe AttachmentReference2KHR

pattern KHR_DEPTH_STENCIL_RESOLVE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_DEPTH_STENCIL_RESOLVE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_DEPTH_STENCIL_RESOLVE_PROPERTIES_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SUBPASS_DESCRIPTION_DEPTH_STENCIL_RESOLVE_KHR :: a ~ StructureType => a