Graphics.Vulkan.Extensions.VK_KHR_device_group

Documentation

type BindBufferMemoryDeviceGroupInfoKHR = BindBufferMemoryDeviceGroupInfo

type BindImageMemoryDeviceGroupInfoKHR = BindImageMemoryDeviceGroupInfo

type DeviceGroupBindSparseInfoKHR = DeviceGroupBindSparseInfo

type DeviceGroupCommandBufferBeginInfoKHR = DeviceGroupCommandBufferBeginInfo

type DeviceGroupRenderPassBeginInfoKHR = DeviceGroupRenderPassBeginInfo

type DeviceGroupSubmitInfoKHR = DeviceGroupSubmitInfo

type MemoryAllocateFlagsInfoKHR = MemoryAllocateFlagsInfo

cmdDispatchBaseKHR :: CommandBuffer -> Word32 -> Word32 -> Word32 -> Word32 -> Word32 -> Word32 -> IO ()

cmdSetDeviceMaskKHR :: CommandBuffer -> Word32 -> IO ()

getDeviceGroupPeerMemoryFeaturesKHR :: Device -> Word32 -> Word32 -> Word32 -> IO PeerMemoryFeatureFlags

pattern DEPENDENCY_DEVICE_GROUP_BIT_KHR :: VkDependencyFlagBits

pattern IMAGE_CREATE_SPLIT_INSTANCE_BIND_REGIONS_BIT_KHR :: VkImageCreateFlagBits

pattern KHR_DEVICE_GROUP_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_DEVICE_GROUP_SPEC_VERSION :: Integral a => a

pattern MEMORY_ALLOCATE_DEVICE_MASK_BIT_KHR :: VkMemoryAllocateFlagBits

pattern PEER_MEMORY_FEATURE_COPY_DST_BIT_KHR :: VkPeerMemoryFeatureFlagBits

pattern PEER_MEMORY_FEATURE_COPY_SRC_BIT_KHR :: VkPeerMemoryFeatureFlagBits

pattern PEER_MEMORY_FEATURE_GENERIC_DST_BIT_KHR :: VkPeerMemoryFeatureFlagBits

pattern PEER_MEMORY_FEATURE_GENERIC_SRC_BIT_KHR :: VkPeerMemoryFeatureFlagBits

pattern PIPELINE_CREATE_DISPATCH_BASE_KHR :: VkPipelineCreateFlagBits

pattern PIPELINE_CREATE_VIEW_INDEX_FROM_DEVICE_INDEX_BIT_KHR :: VkPipelineCreateFlagBits

pattern STRUCTURE_TYPE_BIND_BUFFER_MEMORY_DEVICE_GROUP_INFO_KHR :: VkStructureType

pattern STRUCTURE_TYPE_BIND_IMAGE_MEMORY_DEVICE_GROUP_INFO_KHR :: VkStructureType

pattern STRUCTURE_TYPE_DEVICE_GROUP_BIND_SPARSE_INFO_KHR :: VkStructureType

pattern STRUCTURE_TYPE_DEVICE_GROUP_COMMAND_BUFFER_BEGIN_INFO_KHR :: VkStructureType

pattern STRUCTURE_TYPE_DEVICE_GROUP_RENDER_PASS_BEGIN_INFO_KHR :: VkStructureType

pattern STRUCTURE_TYPE_DEVICE_GROUP_SUBMIT_INFO_KHR :: VkStructureType

pattern STRUCTURE_TYPE_MEMORY_ALLOCATE_FLAGS_INFO_KHR :: VkStructureType

type PeerMemoryFeatureFlagsKHR = PeerMemoryFeatureFlags

type PeerMemoryFeatureFlagBitsKHR = PeerMemoryFeatureFlagBits

type MemoryAllocateFlagsKHR = MemoryAllocateFlags

type MemoryAllocateFlagBitsKHR = MemoryAllocateFlagBits

type DeviceGroupPresentModeFlagBitsKHR = VkDeviceGroupPresentModeFlagBitsKHR

VkDeviceGroupPresentModeFlagBitsKHR - Bitmask specifying supported device group present modes

See Also

VkDeviceGroupPresentInfoKHR, VkDeviceGroupPresentModeFlagsKHR

type DeviceGroupPresentModeFlagsKHR = DeviceGroupPresentModeFlagBitsKHR

VkDeviceGroupPresentModeFlagsKHR - Bitmask of VkDeviceGroupPresentModeFlagBitsKHR

Description

VkDeviceGroupPresentModeFlagsKHR is a bitmask type for setting a mask of zero or more VkDeviceGroupPresentModeFlagBitsKHR.

See Also

VkDeviceGroupPresentCapabilitiesKHR, VkDeviceGroupPresentModeFlagBitsKHR, VkDeviceGroupSwapchainCreateInfoKHR, vkGetDeviceGroupSurfacePresentModes2EXT, vkGetDeviceGroupSurfacePresentModesKHR

data DeviceGroupPresentCapabilitiesKHR

VkDeviceGroupPresentCapabilitiesKHR - Present capabilities from other physical devices

Description

modes always has VK_DEVICE_GROUP_PRESENT_MODE_LOCAL_BIT_KHR set.

The present mode flags are also used when presenting an image, in VkDeviceGroupPresentInfoKHR::mode.

If a device group only includes a single physical device, then modes must equal VK_DEVICE_GROUP_PRESENT_MODE_LOCAL_BIT_KHR.

Valid Usage (Implicit)

See Also

VkDeviceGroupPresentModeFlagsKHR, VkStructureType, vkGetDeviceGroupPresentCapabilitiesKHR

Constructors

DeviceGroupPresentCapabilitiesKHR

Fields

next :: Maybe SomeVkStruct

presentMask :: Vector Word32

modes :: DeviceGroupPresentModeFlagsKHR

getDeviceGroupPresentCapabilitiesKHR :: Device -> IO DeviceGroupPresentCapabilitiesKHR

vkGetDeviceGroupPresentCapabilitiesKHR - Query present capabilities from other physical devices

Parameters

device is the logical device.

pDeviceGroupPresentCapabilities is a pointer to a structure of type VkDeviceGroupPresentCapabilitiesKHR that is filled with the logical device’s capabilities.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDevice, VkDeviceGroupPresentCapabilitiesKHR

getDeviceGroupSurfacePresentModesKHR :: Device -> SurfaceKHR -> IO DeviceGroupPresentModeFlagsKHR

vkGetDeviceGroupSurfacePresentModesKHR - Query present capabilities for a surface

Parameters

device is the logical device.

surface is the surface.

pModes is a pointer to a value of type VkDeviceGroupPresentModeFlagsKHR that is filled with the supported device group present modes for the surface.

Description

The modes returned by this command are not invariant, and may change in response to the surface being moved, resized, or occluded. These modes must be a subset of the modes returned by vkGetDeviceGroupPresentCapabilitiesKHR.

Valid Usage (Implicit)

device must be a valid VkDevice handle

surface must be a valid VkSurfaceKHR handle

pModes must be a valid pointer to a VkDeviceGroupPresentModeFlagsKHR value

Both of device, and surface must have been created, allocated, or retrieved from the same VkInstance

Host Synchronization

Host access to surface must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

See Also

VkDevice, VkDeviceGroupPresentModeFlagsKHR, VkSurfaceKHR

getPhysicalDevicePresentRectanglesKHR :: PhysicalDevice -> SurfaceKHR -> Word32 -> IO (VkResult, Vector Rect2D)

vkGetPhysicalDevicePresentRectanglesKHR - Query present rectangles for a surface on a physical device

Parameters

physicalDevice is the physical device.

surface is the surface.

pRectCount is a pointer to an integer related to the number of rectangles available or queried, as described below.

pRects is either NULL or a pointer to an array of VkRect2D structures.

Description

If pRects is NULL, then the number of rectangles used when presenting the given surface is returned in pRectCount. Otherwise, pRectCount must point to a variable set by the user to the number of elements in the pRects array, and on return the variable is overwritten with the number of structures actually written to pRects. If the value of pRectCount is less than the number of rectangles, at most pRectCount structures will be written. If pRectCount is smaller than the number of rectangles used for the given surface, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

The values returned by this command are not invariant, and may change in response to the surface being moved, resized, or occluded.

The rectangles returned by this command must not overlap.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

surface must be a valid VkSurfaceKHR handle

pRectCount must be a valid pointer to a uint32_t value

If the value referenced by pRectCount is not 0, and pRects is not NULL, pRects must be a valid pointer to an array of pRectCount VkRect2D structures

Both of physicalDevice, and surface must have been created, allocated, or retrieved from the same VkInstance

Host Synchronization

Host access to surface must be externally synchronized

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkPhysicalDevice, VkRect2D, VkSurfaceKHR

data ImageSwapchainCreateInfoKHR

VkImageSwapchainCreateInfoKHR - Specify that an image will be bound to swapchain memory

Valid Usage

If swapchain is not VK_NULL_HANDLE, the fields of VkImageCreateInfo must match the implied image creation parameters of the swapchain

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMAGE_SWAPCHAIN_CREATE_INFO_KHR

If swapchain is not VK_NULL_HANDLE, swapchain must be a valid VkSwapchainKHR handle

See Also

VkStructureType, VkSwapchainKHR

Constructors

ImageSwapchainCreateInfoKHR

Fields

next :: Maybe SomeVkStruct

swapchain :: SwapchainKHR

data BindImageMemorySwapchainInfoKHR

VkBindImageMemorySwapchainInfoKHR - Structure specifying swapchain image memory to bind to

Description

If swapchain is not NULL, the swapchain and imageIndex are used to determine the memory that the image is bound to, instead of memory and memoryOffset.

Memory can be bound to a swapchain and use the pDeviceIndices or pSplitInstanceBindRegions members of VkBindImageMemoryDeviceGroupInfo.

Valid Usage

imageIndex must be less than the number of images in swapchain

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_BIND_IMAGE_MEMORY_SWAPCHAIN_INFO_KHR

swapchain must be a valid VkSwapchainKHR handle

Host Synchronization

Host access to swapchain must be externally synchronized

See Also

VkStructureType, VkSwapchainKHR

Constructors

BindImageMemorySwapchainInfoKHR

Fields

next :: Maybe SomeVkStruct

swapchain :: SwapchainKHR

imageIndex :: Word32

data AcquireNextImageInfoKHR

VkAcquireNextImageInfoKHR - Structure specifying parameters of the acquire

Description

If vkAcquireNextImageKHR is used, the device mask is considered to include all physical devices in the logical device.

Note

vkAcquireNextImage2KHR signals at most one semaphore, even if the application requests waiting for multiple physical devices to be ready via the deviceMask. However, only a single physical device can wait on that semaphore, since the semaphore becomes unsignaled when the wait succeeds. For other physical devices to wait for the image to be ready, it is necessary for the application to submit semaphore signal operation(s) to that first physical device to signal additional semaphore(s) after the wait succeeds, which the other physical device(s) can wait upon.

Valid Usage

swapchain must not be in the retired state

If semaphore is not VK_NULL_HANDLE it must be unsignaled

If semaphore is not VK_NULL_HANDLE it must not have any uncompleted signal or wait operations pending

If fence is not VK_NULL_HANDLE it must be unsignaled and must not be associated with any other queue command that has not yet completed execution on that queue

semaphore and fence must not both be equal to VK_NULL_HANDLE

deviceMask must be a valid device mask

deviceMask must not be zero

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_ACQUIRE_NEXT_IMAGE_INFO_KHR

pNext must be NULL

swapchain must be a valid VkSwapchainKHR handle

If semaphore is not VK_NULL_HANDLE, semaphore must be a valid VkSemaphore handle

If fence is not VK_NULL_HANDLE, fence must be a valid VkFence handle

Each of fence, semaphore, and swapchain that are valid handles must have been created, allocated, or retrieved from the same VkInstance

Host Synchronization

Host access to swapchain must be externally synchronized

Host access to semaphore must be externally synchronized

Host access to fence must be externally synchronized

See Also

VkFence, VkSemaphore, VkStructureType, VkSwapchainKHR, vkAcquireNextImage2KHR

Constructors

AcquireNextImageInfoKHR

Fields

next :: Maybe SomeVkStruct

swapchain :: SwapchainKHR

timeout :: Word64

semaphore :: Semaphore

fence :: Fence

deviceMask :: Word32

data DeviceGroupPresentInfoKHR

VkDeviceGroupPresentInfoKHR - Mode and mask controlling which physical devices' images are presented

Description

If mode is VK_DEVICE_GROUP_PRESENT_MODE_LOCAL_BIT_KHR, then each element of pDeviceMasks selects which instance of the swapchain image is presented. Each element of pDeviceMasks must have exactly one bit set, and the corresponding physical device must have a presentation engine as reported by VkDeviceGroupPresentCapabilitiesKHR.

If mode is VK_DEVICE_GROUP_PRESENT_MODE_REMOTE_BIT_KHR, then each element of pDeviceMasks selects which instance of the swapchain image is presented. Each element of pDeviceMasks must have exactly one bit set, and some physical device in the logical device must include that bit in its VkDeviceGroupPresentCapabilitiesKHR::presentMask.

If mode is VK_DEVICE_GROUP_PRESENT_MODE_SUM_BIT_KHR, then each element of pDeviceMasks selects which instances of the swapchain image are component-wise summed and the sum of those images is presented. If the sum in any component is outside the representable range, the value of that component is undefined. Each element of pDeviceMasks must have a value for which all set bits are set in one of the elements of VkDeviceGroupPresentCapabilitiesKHR::presentMask.

If mode is VK_DEVICE_GROUP_PRESENT_MODE_LOCAL_MULTI_DEVICE_BIT_KHR, then each element of pDeviceMasks selects which instance(s) of the swapchain images are presented. For each bit set in each element of pDeviceMasks, the corresponding physical device must have a presentation engine as reported by VkDeviceGroupPresentCapabilitiesKHR.

If VkDeviceGroupPresentInfoKHR is not provided or swapchainCount is zero then the masks are considered to be 1. If VkDeviceGroupPresentInfoKHR is not provided, mode is considered to be VK_DEVICE_GROUP_PRESENT_MODE_LOCAL_BIT_KHR.

Valid Usage

swapchainCount must equal 0 or VkPresentInfoKHR::swapchainCount

If mode is VK_DEVICE_GROUP_PRESENT_MODE_LOCAL_BIT_KHR, then each element of pDeviceMasks must have exactly one bit set, and the corresponding element of VkDeviceGroupPresentCapabilitiesKHR::presentMask must be non-zero

If mode is VK_DEVICE_GROUP_PRESENT_MODE_REMOTE_BIT_KHR, then each element of pDeviceMasks must have exactly one bit set, and some physical device in the logical device must include that bit in its VkDeviceGroupPresentCapabilitiesKHR::presentMask.

If mode is VK_DEVICE_GROUP_PRESENT_MODE_SUM_BIT_KHR, then each element of pDeviceMasks must have a value for which all set bits are set in one of the elements of VkDeviceGroupPresentCapabilitiesKHR::presentMask

If mode is VK_DEVICE_GROUP_PRESENT_MODE_LOCAL_MULTI_DEVICE_BIT_KHR, then for each bit set in each element of pDeviceMasks, the corresponding element of VkDeviceGroupPresentCapabilitiesKHR::presentMask must be non-zero

The value of each element of pDeviceMasks must be equal to the device mask passed in VkAcquireNextImageInfoKHR::deviceMask when the image index was last acquired

mode must have exactly one bit set, and that bit must have been included in VkDeviceGroupSwapchainCreateInfoKHR::modes

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DEVICE_GROUP_PRESENT_INFO_KHR

If swapchainCount is not 0, pDeviceMasks must be a valid pointer to an array of swapchainCount uint32_t values

mode must be a valid VkDeviceGroupPresentModeFlagBitsKHR value

See Also

VkDeviceGroupPresentModeFlagBitsKHR, VkStructureType

Constructors

DeviceGroupPresentInfoKHR

Fields

next :: Maybe SomeVkStruct

deviceMasks :: Vector Word32

mode :: DeviceGroupPresentModeFlagBitsKHR

data DeviceGroupSwapchainCreateInfoKHR

VkDeviceGroupSwapchainCreateInfoKHR - Structure specifying parameters of a newly created swapchain object

Description

If this structure is not present, modes is considered to be VK_DEVICE_GROUP_PRESENT_MODE_LOCAL_BIT_KHR.

Valid Usage (Implicit)

See Also

VkDeviceGroupPresentModeFlagsKHR, VkStructureType

Constructors

DeviceGroupSwapchainCreateInfoKHR

Fields

next :: Maybe SomeVkStruct

modes :: DeviceGroupPresentModeFlagsKHR

acquireNextImage2KHR :: Device -> AcquireNextImageInfoKHR -> IO (VkResult, Word32)

vkAcquireNextImage2KHR - Retrieve the index of the next available presentable image

Parameters

device is the device associated with swapchain.

pAcquireInfo is a pointer to a structure of type VkAcquireNextImageInfoKHR containing parameters of the acquire.

pImageIndex is a pointer to a uint32_t that is set to the index of the next image to use.

Valid Usage

If the number of currently acquired images is greater than the difference between the number of images in the swapchain member of pAcquireInfo and the value of VkSurfaceCapabilitiesKHR::minImageCount as returned by a call to vkGetPhysicalDeviceSurfaceCapabilities2KHR with the surface used to create swapchain, the timeout member of pAcquireInfo must not be UINT64_MAX

Valid Usage (Implicit)

device must be a valid VkDevice handle

pAcquireInfo must be a valid pointer to a valid VkAcquireNextImageInfoKHR structure

pImageIndex must be a valid pointer to a uint32_t value

Return Codes

Success

- VK_SUCCESS

VK_TIMEOUT

VK_NOT_READY

VK_SUBOPTIMAL_KHR

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_DEVICE_LOST

VK_ERROR_OUT_OF_DATE_KHR

VK_ERROR_SURFACE_LOST_KHR

VK_ERROR_FULL_SCREEN_EXCLUSIVE_MODE_LOST_EXT

See Also

VkAcquireNextImageInfoKHR, VkDevice

pattern STRUCTURE_TYPE_MEMORY_ALLOCATE_FLAGS_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEVICE_GROUP_RENDER_PASS_BEGIN_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEVICE_GROUP_COMMAND_BUFFER_BEGIN_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEVICE_GROUP_SUBMIT_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEVICE_GROUP_BIND_SPARSE_INFO :: a ~ StructureType => a

pattern PEER_MEMORY_FEATURE_COPY_SRC_BIT :: a ~ PeerMemoryFeatureFlagBits => a

VK_PEER_MEMORY_FEATURE_COPY_SRC_BIT specifies that the memory can be accessed as the source of a vkCmdCopyBuffer, vkCmdCopyImage, vkCmdCopyBufferToImage, or vkCmdCopyImageToBuffer command.

pattern PEER_MEMORY_FEATURE_COPY_DST_BIT :: a ~ PeerMemoryFeatureFlagBits => a

VK_PEER_MEMORY_FEATURE_COPY_DST_BIT specifies that the memory can be accessed as the destination of a vkCmdCopyBuffer, vkCmdCopyImage, vkCmdCopyBufferToImage, or vkCmdCopyImageToBuffer command.

pattern PEER_MEMORY_FEATURE_GENERIC_SRC_BIT :: a ~ PeerMemoryFeatureFlagBits => a

VK_PEER_MEMORY_FEATURE_GENERIC_SRC_BIT specifies that the memory can be read as any memory access type.

pattern PEER_MEMORY_FEATURE_GENERIC_DST_BIT :: a ~ PeerMemoryFeatureFlagBits => a

VK_PEER_MEMORY_FEATURE_GENERIC_DST_BIT specifies that the memory can be written as any memory access type. Shader atomics are considered to be writes.

pattern MEMORY_ALLOCATE_DEVICE_MASK_BIT :: a ~ MemoryAllocateFlagBits => a

VK_MEMORY_ALLOCATE_DEVICE_MASK_BIT specifies that memory will be allocated for the devices in VkMemoryAllocateFlagsInfo::deviceMask.

pattern PIPELINE_CREATE_VIEW_INDEX_FROM_DEVICE_INDEX_BIT :: a ~ PipelineCreateFlagBits => a

pattern PIPELINE_CREATE_DISPATCH_BASE :: a ~ PipelineCreateFlagBits => a

pattern DEPENDENCY_DEVICE_GROUP_BIT :: a ~ DependencyFlagBits => a

pattern STRUCTURE_TYPE_BIND_BUFFER_MEMORY_DEVICE_GROUP_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_BIND_IMAGE_MEMORY_DEVICE_GROUP_INFO :: a ~ StructureType => a

pattern IMAGE_CREATE_SPLIT_INSTANCE_BIND_REGIONS_BIT :: a ~ ImageCreateFlagBits => a

pattern STRUCTURE_TYPE_DEVICE_GROUP_PRESENT_CAPABILITIES_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_IMAGE_SWAPCHAIN_CREATE_INFO_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_BIND_IMAGE_MEMORY_SWAPCHAIN_INFO_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_ACQUIRE_NEXT_IMAGE_INFO_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEVICE_GROUP_PRESENT_INFO_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEVICE_GROUP_SWAPCHAIN_CREATE_INFO_KHR :: a ~ StructureType => a

pattern SWAPCHAIN_CREATE_SPLIT_INSTANCE_BIND_REGIONS_BIT_KHR :: a ~ SwapchainCreateFlagBitsKHR => a