Graphics.Vulkan.Extensions.VK_KHR_display

Documentation

type DisplayKHR = VkDisplayKHR

VkDisplayKHR - Opaque handle to a display object

See Also

VkDisplayPlanePropertiesKHR, VkDisplayPropertiesKHR, vkAcquireXlibDisplayEXT, vkCreateDisplayModeKHR, vkDisplayPowerControlEXT, vkGetDisplayModeProperties2KHR, vkGetDisplayModePropertiesKHR, vkGetDisplayPlaneSupportedDisplaysKHR, vkGetRandROutputDisplayEXT, vkRegisterDisplayEventEXT, vkReleaseDisplayEXT

type DisplayModeCreateFlagsKHR = VkDisplayModeCreateFlagsKHR

withCStructDisplayModeCreateInfoKHR :: DisplayModeCreateInfoKHR -> (VkDisplayModeCreateInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkDisplayModeCreateInfoKHR and marshal a DisplayModeCreateInfoKHR into it. The VkDisplayModeCreateInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructDisplayModeCreateInfoKHR :: VkDisplayModeCreateInfoKHR -> IO DisplayModeCreateInfoKHR

A function to read a VkDisplayModeCreateInfoKHR and all additional structures in the pointer chain into a DisplayModeCreateInfoKHR.

data DisplayModeCreateInfoKHR

VkDisplayModeCreateInfoKHR - Structure specifying parameters of a newly created display mode object

Valid Usage (Implicit)

See Also

VkDisplayModeCreateFlagsKHR, VkDisplayModeParametersKHR, VkStructureType, vkCreateDisplayModeKHR

Constructors

DisplayModeCreateInfoKHR

Fields

next :: Maybe SomeVkStruct

flags :: DisplayModeCreateFlagsKHR

parameters :: DisplayModeParametersKHR

type DisplayModeKHR = VkDisplayModeKHR

VkDisplayModeKHR - Opaque handle to a display mode object

See Also

VkDisplayModePropertiesKHR, VkDisplayPlaneInfo2KHR, VkDisplaySurfaceCreateInfoKHR, vkCreateDisplayModeKHR, vkGetDisplayPlaneCapabilitiesKHR

withCStructDisplayModeParametersKHR :: DisplayModeParametersKHR -> (VkDisplayModeParametersKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkDisplayModeParametersKHR and marshal a DisplayModeParametersKHR into it. The VkDisplayModeParametersKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructDisplayModeParametersKHR :: VkDisplayModeParametersKHR -> IO DisplayModeParametersKHR

A function to read a VkDisplayModeParametersKHR and all additional structures in the pointer chain into a DisplayModeParametersKHR.

data DisplayModeParametersKHR

VkDisplayModeParametersKHR - Structure describing display parameters associated with a display mode

Description

Note

For example, a 60Hz display mode would report a refreshRate of 60,000.

Valid Usage

The width member of visibleRegion must be greater than 0

The height member of visibleRegion must be greater than 0

refreshRate must be greater than 0

See Also

VkDisplayModeCreateInfoKHR, VkDisplayModePropertiesKHR, VkExtent2D

Constructors

DisplayModeParametersKHR

Fields

visibleRegion :: Extent2D

refreshRate :: Word32

withCStructDisplayModePropertiesKHR :: DisplayModePropertiesKHR -> (VkDisplayModePropertiesKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkDisplayModePropertiesKHR and marshal a DisplayModePropertiesKHR into it. The VkDisplayModePropertiesKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructDisplayModePropertiesKHR :: VkDisplayModePropertiesKHR -> IO DisplayModePropertiesKHR

A function to read a VkDisplayModePropertiesKHR and all additional structures in the pointer chain into a DisplayModePropertiesKHR.

data DisplayModePropertiesKHR

VkDisplayModePropertiesKHR - Structure describing display mode properties

See Also

VkDisplayModeKHR, VkDisplayModeParametersKHR, VkDisplayModeProperties2KHR, vkGetDisplayModePropertiesKHR

Constructors

DisplayModePropertiesKHR

Fields

displayMode :: DisplayModeKHR

parameters :: DisplayModeParametersKHR

type DisplayPlaneAlphaFlagBitsKHR = VkDisplayPlaneAlphaFlagBitsKHR

VkDisplayPlaneAlphaFlagBitsKHR - Alpha blending type

See Also

VkDisplayPlaneAlphaFlagsKHR, VkDisplaySurfaceCreateInfoKHR

pattern DISPLAY_PLANE_ALPHA_OPAQUE_BIT_KHR :: a ~ DisplayPlaneAlphaFlagBitsKHR => a

VK_DISPLAY_PLANE_ALPHA_OPAQUE_BIT_KHR specifies that the source image will be treated as opaque.

pattern DISPLAY_PLANE_ALPHA_GLOBAL_BIT_KHR :: a ~ DisplayPlaneAlphaFlagBitsKHR => a

VK_DISPLAY_PLANE_ALPHA_GLOBAL_BIT_KHR specifies that a global alpha value must be specified that will be applied to all pixels in the source image.

pattern DISPLAY_PLANE_ALPHA_PER_PIXEL_BIT_KHR :: a ~ DisplayPlaneAlphaFlagBitsKHR => a

VK_DISPLAY_PLANE_ALPHA_PER_PIXEL_BIT_KHR specifies that the alpha value will be determined by the alpha channel of the source image’s pixels. If the source format contains no alpha values, no blending will be applied. The source alpha values are not premultiplied into the source image’s other color channels.

pattern DISPLAY_PLANE_ALPHA_PER_PIXEL_PREMULTIPLIED_BIT_KHR :: a ~ DisplayPlaneAlphaFlagBitsKHR => a

VK_DISPLAY_PLANE_ALPHA_PER_PIXEL_PREMULTIPLIED_BIT_KHR is equivalent to VK_DISPLAY_PLANE_ALPHA_PER_PIXEL_BIT_KHR, except the source alpha values are assumed to be premultiplied into the source image’s other color channels.

type DisplayPlaneAlphaFlagsKHR = DisplayPlaneAlphaFlagBitsKHR

VkDisplayPlaneAlphaFlagsKHR - Bitmask of VkDisplayPlaneAlphaFlagBitsKHR

Description

VkDisplayPlaneAlphaFlagsKHR is a bitmask type for setting a mask of zero or more VkDisplayPlaneAlphaFlagBitsKHR.

See Also

VkDisplayPlaneAlphaFlagBitsKHR, VkDisplayPlaneCapabilitiesKHR

withCStructDisplayPlaneCapabilitiesKHR :: DisplayPlaneCapabilitiesKHR -> (VkDisplayPlaneCapabilitiesKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkDisplayPlaneCapabilitiesKHR and marshal a DisplayPlaneCapabilitiesKHR into it. The VkDisplayPlaneCapabilitiesKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructDisplayPlaneCapabilitiesKHR :: VkDisplayPlaneCapabilitiesKHR -> IO DisplayPlaneCapabilitiesKHR

A function to read a VkDisplayPlaneCapabilitiesKHR and all additional structures in the pointer chain into a DisplayPlaneCapabilitiesKHR.

data DisplayPlaneCapabilitiesKHR

VkDisplayPlaneCapabilitiesKHR - Structure describing capabilities of a mode and plane combination

Description

The minimum and maximum position and extent fields describe the implementation limits, if any, as they apply to the specified display mode and plane. Vendors may support displaying a subset of a swapchain’s presentable images on the specified display plane. This is expressed by returning minSrcPosition, maxSrcPosition, minSrcExtent, and maxSrcExtent values that indicate a range of possible positions and sizes may be used to specify the region within the presentable images that source pixels will be read from when creating a swapchain on the specified display mode and plane.

Vendors may also support mapping the presentable images’ content to a subset or superset of the visible region in the specified display mode. This is expressed by returning minDstPosition, maxDstPosition, minDstExtent and maxDstExtent values that indicate a range of possible positions and sizes may be used to describe the region within the display mode that the source pixels will be mapped to.

Other vendors may support only a 1-1 mapping between pixels in the presentable images and the display mode. This may be indicated by returning (0,0) for minSrcPosition, maxSrcPosition, minDstPosition, and maxDstPosition, and (display mode width, display mode height) for minSrcExtent, maxSrcExtent, minDstExtent, and maxDstExtent.

These values indicate the limits of the implementation’s individual fields. Not all combinations of values within the offset and extent ranges returned in VkDisplayPlaneCapabilitiesKHR are guaranteed to be supported. Vendors may still fail presentation requests that specify unsupported combinations.

See Also

VkDisplayPlaneAlphaFlagsKHR, VkDisplayPlaneCapabilities2KHR, VkExtent2D, VkOffset2D, vkGetDisplayPlaneCapabilitiesKHR

Constructors

DisplayPlaneCapabilitiesKHR

Fields

supportedAlpha :: DisplayPlaneAlphaFlagsKHR

minSrcPosition :: Offset2D

maxSrcPosition :: Offset2D

minSrcExtent :: Extent2D

maxSrcExtent :: Extent2D

minDstPosition :: Offset2D

maxDstPosition :: Offset2D

minDstExtent :: Extent2D

maxDstExtent :: Extent2D

withCStructDisplayPlanePropertiesKHR :: DisplayPlanePropertiesKHR -> (VkDisplayPlanePropertiesKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkDisplayPlanePropertiesKHR and marshal a DisplayPlanePropertiesKHR into it. The VkDisplayPlanePropertiesKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructDisplayPlanePropertiesKHR :: VkDisplayPlanePropertiesKHR -> IO DisplayPlanePropertiesKHR

A function to read a VkDisplayPlanePropertiesKHR and all additional structures in the pointer chain into a DisplayPlanePropertiesKHR.

data DisplayPlanePropertiesKHR

VkDisplayPlanePropertiesKHR - Structure describing display plane properties

See Also

VkDisplayKHR, VkDisplayPlaneProperties2KHR, vkGetPhysicalDeviceDisplayPlanePropertiesKHR

Constructors

DisplayPlanePropertiesKHR

Fields

currentDisplay :: DisplayKHR

currentStackIndex :: Word32

withCStructDisplayPropertiesKHR :: DisplayPropertiesKHR -> (VkDisplayPropertiesKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkDisplayPropertiesKHR and marshal a DisplayPropertiesKHR into it. The VkDisplayPropertiesKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructDisplayPropertiesKHR :: VkDisplayPropertiesKHR -> IO DisplayPropertiesKHR

A function to read a VkDisplayPropertiesKHR and all additional structures in the pointer chain into a DisplayPropertiesKHR.

data DisplayPropertiesKHR

VkDisplayPropertiesKHR - Structure describing an available display device

Description

Note

For devices which have no natural value to return here, implementations should return the maximum resolution supported.

Note

Persistent presents may have higher latency, and may use less power when the screen content is updated infrequently, or when only a portion of the screen needs to be updated in most frames.

See Also

VkBool32, VkDisplayKHR, VkDisplayProperties2KHR, VkExtent2D, VkSurfaceTransformFlagsKHR, vkGetPhysicalDeviceDisplayPropertiesKHR

Constructors

DisplayPropertiesKHR

Fields

display :: DisplayKHR

displayName :: ByteString

physicalDimensions :: Extent2D

physicalResolution :: Extent2D

supportedTransforms :: SurfaceTransformFlagsKHR

planeReorderPossible :: Bool

persistentContent :: Bool

type DisplaySurfaceCreateFlagsKHR = VkDisplaySurfaceCreateFlagsKHR

withCStructDisplaySurfaceCreateInfoKHR :: DisplaySurfaceCreateInfoKHR -> (VkDisplaySurfaceCreateInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkDisplaySurfaceCreateInfoKHR and marshal a DisplaySurfaceCreateInfoKHR into it. The VkDisplaySurfaceCreateInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructDisplaySurfaceCreateInfoKHR :: VkDisplaySurfaceCreateInfoKHR -> IO DisplaySurfaceCreateInfoKHR

A function to read a VkDisplaySurfaceCreateInfoKHR and all additional structures in the pointer chain into a DisplaySurfaceCreateInfoKHR.

data DisplaySurfaceCreateInfoKHR

VkDisplaySurfaceCreateInfoKHR - Structure specifying parameters of a newly created display plane surface object

Description

Note

Creating a display surface must not modify the state of the displays, planes, or other resources it names. For example, it must not apply the specified mode to be set on the associated display. Application of display configuration occurs as a side effect of presenting to a display surface.

Valid Usage

planeIndex must be less than the number of display planes supported by the device as determined by calling vkGetPhysicalDeviceDisplayPlanePropertiesKHR

If the planeReorderPossible member of the VkDisplayPropertiesKHR structure returned by vkGetPhysicalDeviceDisplayPropertiesKHR for the display corresponding to displayMode is VK_TRUE then planeStackIndex must be less than the number of display planes supported by the device as determined by calling vkGetPhysicalDeviceDisplayPlanePropertiesKHR; otherwise planeStackIndex must equal the currentStackIndex member of VkDisplayPlanePropertiesKHR returned by vkGetPhysicalDeviceDisplayPlanePropertiesKHR for the display plane corresponding to displayMode

If alphaMode is VK_DISPLAY_PLANE_ALPHA_GLOBAL_BIT_KHR then globalAlpha must be between 0 and 1, inclusive

alphaMode must be 0 or one of the bits present in the supportedAlpha member of VkDisplayPlaneCapabilitiesKHR returned by vkGetDisplayPlaneCapabilitiesKHR for the display plane corresponding to displayMode

The width and height members of imageExtent must be less than the maxImageDimensions2D member of VkPhysicalDeviceLimits

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DISPLAY_SURFACE_CREATE_INFO_KHR

pNext must be NULL

flags must be 0

displayMode must be a valid VkDisplayModeKHR handle

transform must be a valid VkSurfaceTransformFlagBitsKHR value

alphaMode must be a valid VkDisplayPlaneAlphaFlagBitsKHR value

See Also

VkDisplayModeKHR, VkDisplayPlaneAlphaFlagBitsKHR, VkDisplaySurfaceCreateFlagsKHR, VkExtent2D, VkStructureType, VkSurfaceTransformFlagBitsKHR, vkCreateDisplayPlaneSurfaceKHR

Constructors

DisplaySurfaceCreateInfoKHR

Fields

next :: Maybe SomeVkStruct

flags :: DisplaySurfaceCreateFlagsKHR

displayMode :: DisplayModeKHR

planeIndex :: Word32

planeStackIndex :: Word32

transform :: SurfaceTransformFlagBitsKHR

globalAlpha :: CFloat

alphaMode :: DisplayPlaneAlphaFlagBitsKHR

imageExtent :: Extent2D

createDisplayModeKHR :: PhysicalDevice -> DisplayKHR -> DisplayModeCreateInfoKHR -> Maybe AllocationCallbacks -> IO DisplayModeKHR

vkCreateDisplayModeKHR - Create a display mode

Parameters

physicalDevice is the physical device associated with display.

display is the display to create an additional mode for.

pCreateInfo is a VkDisplayModeCreateInfoKHR structure describing the new mode to create.

pAllocator is the allocator used for host memory allocated for the display mode object when there is no more specific allocator available (see Memory Allocation).

pMode returns the handle of the mode created.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

display must be a valid VkDisplayKHR handle

pCreateInfo must be a valid pointer to a valid VkDisplayModeCreateInfoKHR structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pMode must be a valid pointer to a VkDisplayModeKHR handle

Host Synchronization

Host access to display must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_INITIALIZATION_FAILED

See Also

VkAllocationCallbacks, VkDisplayKHR, VkDisplayModeCreateInfoKHR, VkDisplayModeKHR, VkPhysicalDevice

createDisplayPlaneSurfaceKHR :: Instance -> DisplaySurfaceCreateInfoKHR -> Maybe AllocationCallbacks -> IO SurfaceKHR

vkCreateDisplayPlaneSurfaceKHR - Create a VkSurfaceKHR structure representing a display plane and mode

Parameters

instance is the instance corresponding to the physical device the targeted display is on.

pCreateInfo is a pointer to an instance of the VkDisplaySurfaceCreateInfoKHR structure specifying which mode, plane, and other parameters to use, as described below.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface is returned.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkDisplaySurfaceCreateInfoKHR structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDisplaySurfaceCreateInfoKHR, VkInstance, VkSurfaceKHR

getNumDisplayModePropertiesKHR :: PhysicalDevice -> DisplayKHR -> IO (VkResult, Word32)

vkGetDisplayModePropertiesKHR - Query the set of mode properties supported by the display

Parameters

physicalDevice is the physical device associated with display.

display is the display to query.

pPropertyCount is a pointer to an integer related to the number of display modes available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayModePropertiesKHR structures.

Description

If pProperties is NULL, then the number of display modes available on the specified display for physicalDevice is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If the value of pPropertyCount is less than the number of display modes for physicalDevice, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of display modes available on the specified display for physicalDevice, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

display must be a valid VkDisplayKHR handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayModePropertiesKHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayKHR, VkDisplayModePropertiesKHR, VkPhysicalDevice

getDisplayModePropertiesKHR :: PhysicalDevice -> DisplayKHR -> Word32 -> IO (VkResult, Vector DisplayModePropertiesKHR)

vkGetDisplayModePropertiesKHR - Query the set of mode properties supported by the display

Parameters

physicalDevice is the physical device associated with display.

display is the display to query.

pPropertyCount is a pointer to an integer related to the number of display modes available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayModePropertiesKHR structures.

Description

If pProperties is NULL, then the number of display modes available on the specified display for physicalDevice is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If the value of pPropertyCount is less than the number of display modes for physicalDevice, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of display modes available on the specified display for physicalDevice, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

display must be a valid VkDisplayKHR handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayModePropertiesKHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayKHR, VkDisplayModePropertiesKHR, VkPhysicalDevice

getAllDisplayModePropertiesKHR :: PhysicalDevice -> DisplayKHR -> IO (Vector DisplayModePropertiesKHR)

Returns all the values available from getDisplayModePropertiesKHR.

getDisplayPlaneCapabilitiesKHR :: PhysicalDevice -> DisplayModeKHR -> Word32 -> IO DisplayPlaneCapabilitiesKHR

vkGetDisplayPlaneCapabilitiesKHR - Query capabilities of a mode and plane combination

Parameters

physicalDevice is the physical device associated with display

mode is the display mode the application intends to program when using the specified plane. Note this parameter also implicitly specifies a display.

planeIndex is the plane which the application intends to use with the display, and is less than the number of display planes supported by the device.

pCapabilities is a pointer to a VkDisplayPlaneCapabilitiesKHR structure in which the capabilities are returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

mode must be a valid VkDisplayModeKHR handle

pCapabilities must be a valid pointer to a VkDisplayPlaneCapabilitiesKHR structure

Host Synchronization

Host access to mode must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayModeKHR, VkDisplayPlaneCapabilitiesKHR, VkPhysicalDevice

getNumDisplayPlaneSupportedDisplaysKHR :: PhysicalDevice -> Word32 -> IO (VkResult, Word32)

vkGetDisplayPlaneSupportedDisplaysKHR - Query the list of displays a plane supports

Parameters

physicalDevice is a physical device.

planeIndex is the plane which the application wishes to use, and must be in the range [0, physical device plane count - 1].

pDisplayCount is a pointer to an integer related to the number of displays available or queried, as described below.

pDisplays is either NULL or a pointer to an array of VkDisplayKHR handles.

Description

If pDisplays is NULL, then the number of displays usable with the specified planeIndex for physicalDevice is returned in pDisplayCount. Otherwise, pDisplayCount must point to a variable set by the user to the number of elements in the pDisplays array, and on return the variable is overwritten with the number of handles actually written to pDisplays. If the value of pDisplayCount is less than the number of display planes for physicalDevice, at most pDisplayCount handles will be written. If pDisplayCount is smaller than the number of displays usable with the specified planeIndex for physicalDevice, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage

planeIndex must be less than the number of display planes supported by the device as determined by calling vkGetPhysicalDeviceDisplayPlanePropertiesKHR

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pDisplayCount must be a valid pointer to a uint32_t value

If the value referenced by pDisplayCount is not 0, and pDisplays is not NULL, pDisplays must be a valid pointer to an array of pDisplayCount VkDisplayKHR handles

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayKHR, VkPhysicalDevice

getDisplayPlaneSupportedDisplaysKHR :: PhysicalDevice -> Word32 -> Word32 -> IO (VkResult, Vector DisplayKHR)

vkGetDisplayPlaneSupportedDisplaysKHR - Query the list of displays a plane supports

Parameters

physicalDevice is a physical device.

planeIndex is the plane which the application wishes to use, and must be in the range [0, physical device plane count - 1].

pDisplayCount is a pointer to an integer related to the number of displays available or queried, as described below.

pDisplays is either NULL or a pointer to an array of VkDisplayKHR handles.

Description

If pDisplays is NULL, then the number of displays usable with the specified planeIndex for physicalDevice is returned in pDisplayCount. Otherwise, pDisplayCount must point to a variable set by the user to the number of elements in the pDisplays array, and on return the variable is overwritten with the number of handles actually written to pDisplays. If the value of pDisplayCount is less than the number of display planes for physicalDevice, at most pDisplayCount handles will be written. If pDisplayCount is smaller than the number of displays usable with the specified planeIndex for physicalDevice, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage

planeIndex must be less than the number of display planes supported by the device as determined by calling vkGetPhysicalDeviceDisplayPlanePropertiesKHR

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pDisplayCount must be a valid pointer to a uint32_t value

If the value referenced by pDisplayCount is not 0, and pDisplays is not NULL, pDisplays must be a valid pointer to an array of pDisplayCount VkDisplayKHR handles

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayKHR, VkPhysicalDevice

getAllDisplayPlaneSupportedDisplaysKHR :: PhysicalDevice -> Word32 -> IO (Vector DisplayKHR)

Returns all the values available from getDisplayPlaneSupportedDisplaysKHR.

getNumPhysicalDeviceDisplayPlanePropertiesKHR :: PhysicalDevice -> IO (VkResult, Word32)

vkGetPhysicalDeviceDisplayPlanePropertiesKHR - Query the plane properties

Parameters

physicalDevice is a physical device.

pPropertyCount is a pointer to an integer related to the number of display planes available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayPlanePropertiesKHR structures.

Description

If pProperties is NULL, then the number of display planes available for physicalDevice is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If the value of pPropertyCount is less than the number of display planes for physicalDevice, at most pPropertyCount structures will be written.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayPlanePropertiesKHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayPlanePropertiesKHR, VkPhysicalDevice

getPhysicalDeviceDisplayPlanePropertiesKHR :: PhysicalDevice -> Word32 -> IO (VkResult, Vector DisplayPlanePropertiesKHR)

vkGetPhysicalDeviceDisplayPlanePropertiesKHR - Query the plane properties

Parameters

physicalDevice is a physical device.

pPropertyCount is a pointer to an integer related to the number of display planes available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayPlanePropertiesKHR structures.

Description

If pProperties is NULL, then the number of display planes available for physicalDevice is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If the value of pPropertyCount is less than the number of display planes for physicalDevice, at most pPropertyCount structures will be written.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayPlanePropertiesKHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayPlanePropertiesKHR, VkPhysicalDevice

getAllPhysicalDeviceDisplayPlanePropertiesKHR :: PhysicalDevice -> IO (Vector DisplayPlanePropertiesKHR)

Returns all the values available from getPhysicalDeviceDisplayPlanePropertiesKHR.

getNumPhysicalDeviceDisplayPropertiesKHR :: PhysicalDevice -> IO (VkResult, Word32)

vkGetPhysicalDeviceDisplayPropertiesKHR - Query information about the available displays

Parameters

physicalDevice is a physical device.

pPropertyCount is a pointer to an integer related to the number of display devices available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayPropertiesKHR structures.

Description

If pProperties is NULL, then the number of display devices available for physicalDevice is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If the value of pPropertyCount is less than the number of display devices for physicalDevice, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of display devices available for physicalDevice, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayPropertiesKHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayPropertiesKHR, VkPhysicalDevice

getPhysicalDeviceDisplayPropertiesKHR :: PhysicalDevice -> Word32 -> IO (VkResult, Vector DisplayPropertiesKHR)

vkGetPhysicalDeviceDisplayPropertiesKHR - Query information about the available displays

Parameters

physicalDevice is a physical device.

pPropertyCount is a pointer to an integer related to the number of display devices available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayPropertiesKHR structures.

Description

If pProperties is NULL, then the number of display devices available for physicalDevice is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If the value of pPropertyCount is less than the number of display devices for physicalDevice, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of display devices available for physicalDevice, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayPropertiesKHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayPropertiesKHR, VkPhysicalDevice

getAllPhysicalDeviceDisplayPropertiesKHR :: PhysicalDevice -> IO (Vector DisplayPropertiesKHR)

Returns all the values available from getPhysicalDeviceDisplayPropertiesKHR.

pattern KHR_DISPLAY_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_DISPLAY_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_DISPLAY_MODE_CREATE_INFO_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DISPLAY_SURFACE_CREATE_INFO_KHR :: a ~ StructureType => a

pattern OBJECT_TYPE_DISPLAY_KHR :: a ~ ObjectType => a

pattern OBJECT_TYPE_DISPLAY_MODE_KHR :: a ~ ObjectType => a