Graphics.Vulkan.Extensions.VK_KHR_display_swapchain

Documentation

withCStructDisplayPresentInfoKHR :: DisplayPresentInfoKHR -> (VkDisplayPresentInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkDisplayPresentInfoKHR and marshal a DisplayPresentInfoKHR into it. The VkDisplayPresentInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructDisplayPresentInfoKHR :: VkDisplayPresentInfoKHR -> IO DisplayPresentInfoKHR

A function to read a VkDisplayPresentInfoKHR and all additional structures in the pointer chain into a DisplayPresentInfoKHR.

data DisplayPresentInfoKHR

VkDisplayPresentInfoKHR - Structure describing parameters of a queue presentation to a swapchain

Description

If the extent of the srcRect and dstRect are not equal, the presented pixels will be scaled accordingly.

Valid Usage

srcRect must specify a rectangular region that is a subset of the image being presented

dstRect must specify a rectangular region that is a subset of the visibleRegion parameter of the display mode the swapchain being presented uses

If the persistentContent member of the VkDisplayPropertiesKHR structure returned by vkGetPhysicalDeviceDisplayPropertiesKHR for the display the present operation targets then persistent must be VK_FALSE

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DISPLAY_PRESENT_INFO_KHR

See Also

VkBool32, VkRect2D, VkStructureType

Constructors

DisplayPresentInfoKHR

Fields

next :: Maybe SomeVkStruct

srcRect :: Rect2D

dstRect :: Rect2D

persistent :: Bool

createSharedSwapchainsKHR :: Device -> Vector SwapchainCreateInfoKHR -> Maybe AllocationCallbacks -> IO (Vector SwapchainKHR)

vkCreateSharedSwapchainsKHR - Create multiple swapchains that share presentable images

Parameters

device is the device to create the swapchains for.

swapchainCount is the number of swapchains to create.

pCreateInfos is a pointer to an array of VkSwapchainCreateInfoKHR structures specifying the parameters of the created swapchains.

pAllocator is the allocator used for host memory allocated for the swapchain objects when there is no more specific allocator available (see Memory Allocation).

pSwapchains is a pointer to an array of VkSwapchainKHR handles in which the created swapchain objects will be returned.

Description

vkCreateSharedSwapchainsKHR is similar to vkCreateSwapchainKHR, except that it takes an array of VkSwapchainCreateInfoKHR structures, and returns an array of swapchain objects.

The swapchain creation parameters that affect the properties and number of presentable images must match between all the swapchains. If the displays used by any of the swapchains do not use the same presentable image layout or are incompatible in a way that prevents sharing images, swapchain creation will fail with the result code VK_ERROR_INCOMPATIBLE_DISPLAY_KHR. If any error occurs, no swapchains will be created. Images presented to multiple swapchains must be re-acquired from all of them before transitioning away from VK_IMAGE_LAYOUT_PRESENT_SRC_KHR. After destroying one or more of the swapchains, the remaining swapchains and the presentable images can continue to be used.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfos must be a valid pointer to an array of swapchainCount valid VkSwapchainCreateInfoKHR structures

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSwapchains must be a valid pointer to an array of swapchainCount VkSwapchainKHR handles

swapchainCount must be greater than 0

Host Synchronization

Host access to pCreateInfos[].surface must be externally synchronized

Host access to pCreateInfos[].oldSwapchain must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_INCOMPATIBLE_DISPLAY_KHR

VK_ERROR_DEVICE_LOST

VK_ERROR_SURFACE_LOST_KHR

See Also

VkAllocationCallbacks, VkDevice, VkSwapchainCreateInfoKHR, VkSwapchainKHR

pattern KHR_DISPLAY_SWAPCHAIN_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_DISPLAY_SWAPCHAIN_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_DISPLAY_PRESENT_INFO_KHR :: a ~ StructureType => a

pattern ERROR_INCOMPATIBLE_DISPLAY_KHR :: a ~ Result => a