Graphics.Vulkan.Extensions.VK_KHR_driver_properties

Documentation

withCStructConformanceVersionKHR :: ConformanceVersionKHR -> (VkConformanceVersionKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkConformanceVersionKHR and marshal a ConformanceVersionKHR into it. The VkConformanceVersionKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructConformanceVersionKHR :: VkConformanceVersionKHR -> IO ConformanceVersionKHR

A function to read a VkConformanceVersionKHR and all additional structures in the pointer chain into a ConformanceVersionKHR.

data ConformanceVersionKHR

VkConformanceVersionKHR - Structure containing the conformance test suite version the implementation is compliant with

See Also

VkPhysicalDeviceDriverPropertiesKHR

Constructors

ConformanceVersionKHR

Fields

major :: Word8

minor :: Word8

subminor :: Word8

patch :: Word8

type DriverIdKHR = VkDriverIdKHR

VkDriverIdKHR - Khronos driver IDs

Description

Note

Khronos driver IDs may be allocated by vendors at any time. There may be multiple driver IDs for the same vendor, representing different drivers (for e.g. different platforms, proprietary or open source, etc.). Only the latest canonical versions of this Specification, of the corresponding vk.xml API Registry, and of the corresponding vulkan_core.h header file must contain all reserved Khronos driver IDs.

Only driver IDs registered with Khronos are given symbolic names. There may be unregistered driver IDs returned.

See Also

VkPhysicalDeviceDriverPropertiesKHR

pattern DRIVER_ID_AMD_PROPRIETARY_KHR :: a ~ DriverIdKHR => a

pattern DRIVER_ID_AMD_OPEN_SOURCE_KHR :: a ~ DriverIdKHR => a

pattern DRIVER_ID_MESA_RADV_KHR :: a ~ DriverIdKHR => a

pattern DRIVER_ID_NVIDIA_PROPRIETARY_KHR :: a ~ DriverIdKHR => a

pattern DRIVER_ID_INTEL_PROPRIETARY_WINDOWS_KHR :: a ~ DriverIdKHR => a

pattern DRIVER_ID_INTEL_OPEN_SOURCE_MESA_KHR :: a ~ DriverIdKHR => a

pattern DRIVER_ID_IMAGINATION_PROPRIETARY_KHR :: a ~ DriverIdKHR => a

pattern DRIVER_ID_QUALCOMM_PROPRIETARY_KHR :: a ~ DriverIdKHR => a

pattern DRIVER_ID_ARM_PROPRIETARY_KHR :: a ~ DriverIdKHR => a

pattern DRIVER_ID_GOOGLE_PASTEL_KHR :: a ~ DriverIdKHR => a

pattern DRIVER_ID_GGP_PROPRIETARY_KHR :: a ~ DriverIdKHR => a

withCStructPhysicalDeviceDriverPropertiesKHR :: PhysicalDeviceDriverPropertiesKHR -> (VkPhysicalDeviceDriverPropertiesKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceDriverPropertiesKHR and marshal a PhysicalDeviceDriverPropertiesKHR into it. The VkPhysicalDeviceDriverPropertiesKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceDriverPropertiesKHR :: VkPhysicalDeviceDriverPropertiesKHR -> IO PhysicalDeviceDriverPropertiesKHR

A function to read a VkPhysicalDeviceDriverPropertiesKHR and all additional structures in the pointer chain into a PhysicalDeviceDriverPropertiesKHR.

data PhysicalDeviceDriverPropertiesKHR

VkPhysicalDeviceDriverPropertiesKHR - Structure containing driver identification information

Description

driverID must be immutable for a given driver across instances, processes, driver versions, and system reboots.

Valid Usage (Implicit)

See Also

VkConformanceVersionKHR, VkDriverIdKHR, VkStructureType

Constructors

PhysicalDeviceDriverPropertiesKHR

Fields

next :: Maybe SomeVkStruct

driverID :: DriverIdKHR

driverName :: ByteString

driverInfo :: ByteString

conformanceVersion :: ConformanceVersionKHR

pattern KHR_DRIVER_PROPERTIES_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_DRIVER_PROPERTIES_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_DRIVER_PROPERTIES_KHR :: a ~ StructureType => a