Graphics.Vulkan.Extensions.VK_KHR_external_fence

Documentation

type ExportFenceCreateInfoKHR = ExportFenceCreateInfo

pattern FENCE_IMPORT_TEMPORARY_BIT_KHR :: VkFenceImportFlagBits

pattern KHR_EXTERNAL_FENCE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_EXTERNAL_FENCE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_EXPORT_FENCE_CREATE_INFO_KHR :: VkStructureType

type FenceImportFlagsKHR = FenceImportFlags

type FenceImportFlagBitsKHR = FenceImportFlagBits

pattern STRUCTURE_TYPE_EXPORT_FENCE_CREATE_INFO :: a ~ StructureType => a

pattern FENCE_IMPORT_TEMPORARY_BIT :: a ~ FenceImportFlagBits => a

VK_FENCE_IMPORT_TEMPORARY_BIT specifies that the fence payload will be imported only temporarily, as described in Importing Fence Payloads, regardless of the permanence of handleType.