Graphics.Vulkan.Extensions.VK_KHR_external_fence_capabilities

Documentation

type ExternalFencePropertiesKHR = ExternalFenceProperties

type PhysicalDeviceExternalFenceInfoKHR = PhysicalDeviceExternalFenceInfo

getPhysicalDeviceExternalFencePropertiesKHR :: PhysicalDevice -> PhysicalDeviceExternalFenceInfo -> IO ExternalFenceProperties

pattern EXTERNAL_FENCE_FEATURE_EXPORTABLE_BIT_KHR :: VkExternalFenceFeatureFlagBits

pattern EXTERNAL_FENCE_FEATURE_IMPORTABLE_BIT_KHR :: VkExternalFenceFeatureFlagBits

pattern EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_FD_BIT_KHR :: VkExternalFenceHandleTypeFlagBits

pattern EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_BIT_KHR :: VkExternalFenceHandleTypeFlagBits

pattern EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_KHR :: VkExternalFenceHandleTypeFlagBits

pattern EXTERNAL_FENCE_HANDLE_TYPE_SYNC_FD_BIT_KHR :: VkExternalFenceHandleTypeFlagBits

pattern KHR_EXTERNAL_FENCE_CAPABILITIES_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_EXTERNAL_FENCE_CAPABILITIES_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_EXTERNAL_FENCE_PROPERTIES_KHR :: VkStructureType

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_FENCE_INFO_KHR :: VkStructureType

type ExternalFenceHandleTypeFlagsKHR = ExternalFenceHandleTypeFlags

type ExternalFenceHandleTypeFlagBitsKHR = ExternalFenceHandleTypeFlagBits

type ExternalFenceFeatureFlagsKHR = ExternalFenceFeatureFlags

type ExternalFenceFeatureFlagBitsKHR = ExternalFenceFeatureFlagBits

type PhysicalDeviceIDPropertiesKHR = PhysicalDeviceIDProperties

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_FENCE_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_EXTERNAL_FENCE_PROPERTIES :: a ~ StructureType => a

pattern EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_FD_BIT :: a ~ ExternalFenceHandleTypeFlagBits => a

VK_EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_FD_BIT specifies a POSIX file descriptor handle that has only limited valid usage outside of Vulkan and other compatible APIs. It must be compatible with the POSIX system calls dup, dup2, close, and the non-standard system call dup3. Additionally, it must be transportable over a socket using an SCM_RIGHTS control message. It owns a reference to the underlying synchronization primitive represented by its Vulkan fence object.

pattern EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_BIT :: a ~ ExternalFenceHandleTypeFlagBits => a

VK_EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_BIT specifies an NT handle that has only limited valid usage outside of Vulkan and other compatible APIs. It must be compatible with the functions DuplicateHandle, CloseHandle, CompareObjectHandles, GetHandleInformation, and SetHandleInformation. It owns a reference to the underlying synchronization primitive represented by its Vulkan fence object.

pattern EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT :: a ~ ExternalFenceHandleTypeFlagBits => a

VK_EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT specifies a global share handle that has only limited valid usage outside of Vulkan and other compatible APIs. It is not compatible with any native APIs. It does not own a reference to the underlying synchronization primitive represented by its Vulkan fence object, and will therefore become invalid when all Vulkan fence objects associated with it are destroyed.

pattern EXTERNAL_FENCE_HANDLE_TYPE_SYNC_FD_BIT :: a ~ ExternalFenceHandleTypeFlagBits => a

VK_EXTERNAL_FENCE_HANDLE_TYPE_SYNC_FD_BIT specifies a POSIX file descriptor handle to a Linux Sync File or Android Fence. It can be used with any native API accepting a valid sync file or fence as input. It owns a reference to the underlying synchronization primitive associated with the file descriptor. Implementations which support importing this handle type must accept any type of sync or fence FD supported by the native system they are running on.

pattern EXTERNAL_FENCE_FEATURE_EXPORTABLE_BIT :: a ~ ExternalFenceFeatureFlagBits => a

VK_EXTERNAL_FENCE_FEATURE_EXPORTABLE_BIT specifies handles of this type can be exported from Vulkan fence objects.

pattern EXTERNAL_FENCE_FEATURE_IMPORTABLE_BIT :: a ~ ExternalFenceFeatureFlagBits => a

VK_EXTERNAL_FENCE_FEATURE_IMPORTABLE_BIT specifies handles of this type can be imported to Vulkan fence objects.