Graphics.Vulkan.Extensions.VK_KHR_external_memory_win32

Documentation

getMemoryWin32HandleKHR :: Device -> MemoryGetWin32HandleInfoKHR -> IO HANDLE

vkGetMemoryWin32HandleKHR - Get a Windows HANDLE for a memory object

Parameters

Device is the logical device that created the device memory being exported.

pGetWin32HandleInfo is a pointer to a MemoryGetWin32HandleInfoKHR structure containing parameters of the export operation.

pHandle will return the Windows handle representing the underlying resources of the device memory object.

Description

For handle types defined as NT handles, the handles returned by getMemoryWin32HandleKHR are owned by the application. To avoid leaking resources, the application must release ownership of them using the CloseHandle system call when they are no longer needed.

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

Device, MemoryGetWin32HandleInfoKHR

getMemoryWin32HandlePropertiesKHR :: Device -> ExternalMemoryHandleTypeFlagBits -> HANDLE -> IO MemoryWin32HandlePropertiesKHR

vkGetMemoryWin32HandlePropertiesKHR - Get Properties of External Memory Win32 Handles

Parameters

Device is the logical device that will be importing handle.

handleType is the type of the handle handle.

handle is the handle which will be imported.

pMemoryWin32HandleProperties will return properties of handle.

Return Codes

Success

SUCCESS

Failure

ERROR_INVALID_EXTERNAL_HANDLE

See Also

Device, ExternalMemoryHandleTypeFlagBits, MemoryWin32HandlePropertiesKHR

data ImportMemoryWin32HandleInfoKHR

VkImportMemoryWin32HandleInfoKHR - import Win32 memory created on the same physical device

Description

Importing memory objects from Windows handles does not transfer ownership of the handle to the Vulkan implementation. For handle types defined as NT handles, the application must release ownership using the CloseHandle system call when the handle is no longer needed.

Applications can import the same underlying memory into multiple instances of Vulkan, into the same instance from which it was exported, and multiple times into a given Vulkan instance. In all cases, each import operation must create a distinct DeviceMemory object.

Valid Usage

If handleType is not 0, it must be supported for import, as reported by ExternalImageFormatProperties or ExternalBufferProperties.

The memory from which handle was exported, or the memory named by name must have been created on the same underlying physical device as Device.

If handleType is not 0, it must be defined as an NT handle or a global share handle.

If handleType is not EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT, EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_BIT, EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_HEAP_BIT, or EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_RESOURCE_BIT, name must be NULL.

If handleType is not 0 and handle is NULL, name must name a valid memory resource of the type specified by handleType.

If handleType is not 0 and name is NULL, handle must be a valid handle of the type specified by handleType.

if handle is not NULL, name must be NULL.

If handle is not NULL, it must obey any requirements listed for handleType in external memory handle types compatibility.

If name is not NULL, it must obey any requirements listed for handleType in external memory handle types compatibility.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_IMPORT_MEMORY_WIN32_HANDLE_INFO_KHR

If handleType is not 0, handleType must be a valid ExternalMemoryHandleTypeFlagBits value

See Also

ExternalMemoryHandleTypeFlagBits, StructureType

Constructors

ImportMemoryWin32HandleInfoKHR

Fields

handleType :: ExternalMemoryHandleTypeFlagBits

handleType specifies the type of handle or name.

handle :: HANDLE

handle is the external handle to import, or NULL.

name :: LPCWSTR

name is a null-terminated UTF-16 string naming the underlying memory resource to import, or NULL.

data ExportMemoryWin32HandleInfoKHR

VkExportMemoryWin32HandleInfoKHR - Structure specifying additional attributes of Windows handles exported from a memory

Description

If this structure is not present, or if pAttributes is set to NULL, default security descriptor values will be used, and child processes created by the application will not inherit the handle, as described in the MSDN documentation for “Synchronization Object Security and Access Rights”1. Further, if the structure is not present, the access rights used depend on the handle type.

For handles of the following types:

EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_BIT

The implementation must ensure the access rights allow read and write access to the memory.

For handles of the following types:

EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_HEAP_BIT EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_RESOURCE_BIT

The access rights must be:

GENERIC_ALL

1

https://docs.microsoft.com/en-us/windows/win32/sync/synchronization-object-security-and-access-rights

Valid Usage

If ExportMemoryAllocateInfo::handleTypes does not include EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT, EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_BIT, EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_HEAP_BIT, or EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_RESOURCE_BIT, a ExportMemoryWin32HandleInfoKHR structure must not be included in the pNext chain of MemoryAllocateInfo.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_EXPORT_MEMORY_WIN32_HANDLE_INFO_KHR

If pAttributes is not NULL, pAttributes must be a valid pointer to a valid SECURITY_ATTRIBUTES value

See Also

StructureType

Constructors

ExportMemoryWin32HandleInfoKHR

Fields

attributes :: Ptr SECURITY_ATTRIBUTES

pAttributes is a pointer to a Windows SECURITY_ATTRIBUTES structure specifying security attributes of the handle.

dwAccess :: DWORD

dwAccess is a DWORD specifying access rights of the handle.

name :: LPCWSTR

name is a null-terminated UTF-16 string to associate with the underlying resource referenced by NT handles exported from the created memory.

data MemoryWin32HandlePropertiesKHR

VkMemoryWin32HandlePropertiesKHR - Properties of External Memory Windows Handles

Valid Usage (Implicit)

See Also

StructureType, getMemoryWin32HandlePropertiesKHR

Constructors

MemoryWin32HandlePropertiesKHR

Fields

memoryTypeBits :: Word32

memoryTypeBits is a bitmask containing one bit set for every memory type which the specified windows handle can be imported as.

data MemoryGetWin32HandleInfoKHR

VkMemoryGetWin32HandleInfoKHR - Structure describing a Win32 handle semaphore export operation

Description

The properties of the handle returned depend on the value of handleType. See ExternalMemoryHandleTypeFlagBits for a description of the properties of the defined external memory handle types.

Valid Usage

handleType must have been included in ExportMemoryAllocateInfo::handleTypes when memory was created.

If handleType is defined as an NT handle, getMemoryWin32HandleKHR must be called no more than once for each valid unique combination of memory and handleType.

handleType must be defined as an NT handle or a global share handle.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_MEMORY_GET_WIN32_HANDLE_INFO_KHR

pNext must be NULL

memory must be a valid DeviceMemory handle

handleType must be a valid ExternalMemoryHandleTypeFlagBits value

See Also

DeviceMemory, ExternalMemoryHandleTypeFlagBits, StructureType, getMemoryWin32HandleKHR

Constructors

MemoryGetWin32HandleInfoKHR

Fields

memory :: DeviceMemory

memory is the memory object from which the handle will be exported.

handleType :: ExternalMemoryHandleTypeFlagBits

handleType is the type of handle requested.

type KHR_EXTERNAL_MEMORY_WIN32_SPEC_VERSION = 1

pattern KHR_EXTERNAL_MEMORY_WIN32_SPEC_VERSION :: forall a. Integral a => a

type KHR_EXTERNAL_MEMORY_WIN32_EXTENSION_NAME = "VK_KHR_external_memory_win32"

pattern KHR_EXTERNAL_MEMORY_WIN32_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

type HANDLE = Ptr ()

type DWORD = Word32

type LPCWSTR = Ptr CWchar

data SECURITY_ATTRIBUTES