Graphics.Vulkan.Extensions.VK_KHR_external_semaphore

Documentation

type ExportSemaphoreCreateInfoKHR = ExportSemaphoreCreateInfo

pattern KHR_EXTERNAL_SEMAPHORE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_EXTERNAL_SEMAPHORE_SPEC_VERSION :: Integral a => a

pattern SEMAPHORE_IMPORT_TEMPORARY_BIT_KHR :: VkSemaphoreImportFlagBits

pattern STRUCTURE_TYPE_EXPORT_SEMAPHORE_CREATE_INFO_KHR :: VkStructureType

type SemaphoreImportFlagsKHR = SemaphoreImportFlags

type SemaphoreImportFlagBitsKHR = SemaphoreImportFlagBits

pattern STRUCTURE_TYPE_EXPORT_SEMAPHORE_CREATE_INFO :: a ~ StructureType => a

pattern SEMAPHORE_IMPORT_TEMPORARY_BIT :: a ~ SemaphoreImportFlagBits => a

VK_SEMAPHORE_IMPORT_TEMPORARY_BIT specifies that the semaphore payload will be imported only temporarily, as described in Importing Semaphore Payloads, regardless of the permanence of handleType.