Graphics.Vulkan.Extensions.VK_KHR_external_semaphore_fd

Documentation

getSemaphoreFdKHR :: Device -> SemaphoreGetFdInfoKHR -> IO ("fd" ::: Int32)

vkGetSemaphoreFdKHR - Get a POSIX file descriptor handle for a semaphore

Parameters

Device is the logical device that created the semaphore being exported.

pGetFdInfo is a pointer to a SemaphoreGetFdInfoKHR structure containing parameters of the export operation.

pFd will return the file descriptor representing the semaphore payload.

Description

Each call to getSemaphoreFdKHR must create a new file descriptor and transfer ownership of it to the application. To avoid leaking resources, the application must release ownership of the file descriptor when it is no longer needed.

Note

Ownership can be released in many ways. For example, the application can call close() on the file descriptor, or transfer ownership back to Vulkan by using the file descriptor to import a semaphore payload.

Where supported by the operating system, the implementation must set the file descriptor to be closed automatically when an execve system call is made.

Exporting a file descriptor from a semaphore may have side effects depending on the transference of the specified handle type, as described in Importing Semaphore State.

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

Device, SemaphoreGetFdInfoKHR

importSemaphoreFdKHR :: Device -> ImportSemaphoreFdInfoKHR -> IO ()

vkImportSemaphoreFdKHR - Import a semaphore from a POSIX file descriptor

Parameters

Device is the logical device that created the semaphore.

pImportSemaphoreFdInfo is a pointer to a ImportSemaphoreFdInfoKHR structure specifying the semaphore and import parameters.

Description

Importing a semaphore payload from a file descriptor transfers ownership of the file descriptor from the application to the Vulkan implementation. The application must not perform any operations on the file descriptor after a successful import.

Applications can import the same semaphore payload into multiple instances of Vulkan, into the same instance from which it was exported, and multiple times into a given Vulkan instance.

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_INVALID_EXTERNAL_HANDLE

See Also

Device, ImportSemaphoreFdInfoKHR

data ImportSemaphoreFdInfoKHR

VkImportSemaphoreFdInfoKHR - Structure specifying POSIX file descriptor to import to a semaphore

Description

The handle types supported by handleType are:

Handle Type 	 Transference 	 Permanence Supported

EXTERNAL_SEMAPHORE_HANDLE_TYPE_OPAQUE_FD_BIT 	 Reference 	 Temporary,Permanent

EXTERNAL_SEMAPHORE_HANDLE_TYPE_SYNC_FD_BIT 	 Copy 	 Temporary

Handle Types Supported by ImportSemaphoreFdInfoKHR

Valid Usage

handleType must be a value included in the Handle Types Supported by VkImportSemaphoreFdInfoKHR table.

fd must obey any requirements listed for handleType in external semaphore handle types compatibility.

If handleType is EXTERNAL_SEMAPHORE_HANDLE_TYPE_OPAQUE_FD_BIT, the 'Graphics.Vulkan.Core10.QueueSemaphore.SemaphoreCreateInfo'::'Graphics.Vulkan.Core10.BaseType.Flags' field must match that of the semaphore from which fd was exported.

If handleType is EXTERNAL_SEMAPHORE_HANDLE_TYPE_OPAQUE_FD_BIT, the 'Graphics.Vulkan.Core12.Promoted_From_VK_KHR_timeline_semaphore.SemaphoreTypeCreateInfo'::'Graphics.Vulkan.Core12.Enums.SemaphoreType.SemaphoreType' field must match that of the semaphore from which fd was exported.

If Flags contains SEMAPHORE_IMPORT_TEMPORARY_BIT, the 'Graphics.Vulkan.Core12.Promoted_From_VK_KHR_timeline_semaphore.SemaphoreTypeCreateInfo'::'Graphics.Vulkan.Core12.Enums.SemaphoreType.SemaphoreType' field of the semaphore from which fd was exported must not be SEMAPHORE_TYPE_TIMELINE.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_IMPORT_SEMAPHORE_FD_INFO_KHR

pNext must be NULL

Semaphore must be a valid Semaphore handle

Flags must be a valid combination of SemaphoreImportFlagBits values

handleType must be a valid ExternalSemaphoreHandleTypeFlagBits value

Host Synchronization

Host access to Semaphore must be externally synchronized

See Also

ExternalSemaphoreHandleTypeFlagBits, Semaphore, SemaphoreImportFlags, StructureType, importSemaphoreFdKHR

Constructors

ImportSemaphoreFdInfoKHR

Fields

semaphore :: Semaphore

Semaphore is the semaphore into which the payload will be imported.

flags :: SemaphoreImportFlags

Flags is a bitmask of SemaphoreImportFlagBits specifying additional parameters for the semaphore payload import operation.

handleType :: ExternalSemaphoreHandleTypeFlagBits

handleType specifies the type of fd.

fd :: Int32

fd is the external handle to import.

data SemaphoreGetFdInfoKHR

VkSemaphoreGetFdInfoKHR - Structure describing a POSIX FD semaphore export operation

Description

The properties of the file descriptor returned depend on the value of handleType. See ExternalSemaphoreHandleTypeFlagBits for a description of the properties of the defined external semaphore handle types.

Valid Usage

handleType must have been included in ExportSemaphoreCreateInfo::handleTypes when Semaphore’s current payload was created.

Semaphore must not currently have its payload replaced by an imported payload as described below in Importing Semaphore Payloads unless that imported payload’s handle type was included in ExternalSemaphoreProperties::exportFromImportedHandleTypes for handleType.

If handleType refers to a handle type with copy payload transference semantics, as defined below in Importing Semaphore Payloads, there must be no queue waiting on Semaphore.

If handleType refers to a handle type with copy payload transference semantics, Semaphore must be signaled, or have an associated semaphore signal operation pending execution.

handleType must be defined as a POSIX file descriptor handle.

If handleType refers to a handle type with copy payload transference semantics, Semaphore must have been created with a SemaphoreType of SEMAPHORE_TYPE_BINARY.

If handleType refers to a handle type with copy payload transference semantics, Semaphore must have an associated semaphore signal operation that has been submitted for execution and any semaphore signal operations on which it depends (if any) must have also been submitted for execution.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_SEMAPHORE_GET_FD_INFO_KHR

pNext must be NULL

Semaphore must be a valid Semaphore handle

handleType must be a valid ExternalSemaphoreHandleTypeFlagBits value

See Also

ExternalSemaphoreHandleTypeFlagBits, Semaphore, StructureType, getSemaphoreFdKHR

Constructors

SemaphoreGetFdInfoKHR

Fields

semaphore :: Semaphore

Semaphore is the semaphore from which state will be exported.

handleType :: ExternalSemaphoreHandleTypeFlagBits

handleType is the type of handle requested.

type KHR_EXTERNAL_SEMAPHORE_FD_SPEC_VERSION = 1

pattern KHR_EXTERNAL_SEMAPHORE_FD_SPEC_VERSION :: forall a. Integral a => a

type KHR_EXTERNAL_SEMAPHORE_FD_EXTENSION_NAME = "VK_KHR_external_semaphore_fd"

pattern KHR_EXTERNAL_SEMAPHORE_FD_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a