Graphics.Vulkan.Extensions.VK_KHR_get_display_properties2

Documentation

withCStructDisplayModeProperties2KHR :: DisplayModeProperties2KHR -> (VkDisplayModeProperties2KHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkDisplayModeProperties2KHR and marshal a DisplayModeProperties2KHR into it. The VkDisplayModeProperties2KHR is only valid inside the provided computation and must not be returned out of it.

fromCStructDisplayModeProperties2KHR :: VkDisplayModeProperties2KHR -> IO DisplayModeProperties2KHR

A function to read a VkDisplayModeProperties2KHR and all additional structures in the pointer chain into a DisplayModeProperties2KHR.

data DisplayModeProperties2KHR

VkDisplayModeProperties2KHR - Structure describing an available display mode

Valid Usage (Implicit)

See Also

VkDisplayModePropertiesKHR, VkStructureType, vkGetDisplayModeProperties2KHR

Constructors

DisplayModeProperties2KHR

Fields

next :: Maybe SomeVkStruct

displayModeProperties :: DisplayModePropertiesKHR

withCStructDisplayPlaneCapabilities2KHR :: DisplayPlaneCapabilities2KHR -> (VkDisplayPlaneCapabilities2KHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkDisplayPlaneCapabilities2KHR and marshal a DisplayPlaneCapabilities2KHR into it. The VkDisplayPlaneCapabilities2KHR is only valid inside the provided computation and must not be returned out of it.

fromCStructDisplayPlaneCapabilities2KHR :: VkDisplayPlaneCapabilities2KHR -> IO DisplayPlaneCapabilities2KHR

A function to read a VkDisplayPlaneCapabilities2KHR and all additional structures in the pointer chain into a DisplayPlaneCapabilities2KHR.

data DisplayPlaneCapabilities2KHR

VkDisplayPlaneCapabilities2KHR - Structure describing the capabilities of a mode and plane combination

Valid Usage (Implicit)

See Also

VkDisplayPlaneCapabilitiesKHR, VkStructureType, vkGetDisplayPlaneCapabilities2KHR

Constructors

DisplayPlaneCapabilities2KHR

Fields

next :: Maybe SomeVkStruct

capabilities :: DisplayPlaneCapabilitiesKHR

withCStructDisplayPlaneInfo2KHR :: DisplayPlaneInfo2KHR -> (VkDisplayPlaneInfo2KHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkDisplayPlaneInfo2KHR and marshal a DisplayPlaneInfo2KHR into it. The VkDisplayPlaneInfo2KHR is only valid inside the provided computation and must not be returned out of it.

fromCStructDisplayPlaneInfo2KHR :: VkDisplayPlaneInfo2KHR -> IO DisplayPlaneInfo2KHR

A function to read a VkDisplayPlaneInfo2KHR and all additional structures in the pointer chain into a DisplayPlaneInfo2KHR.

data DisplayPlaneInfo2KHR

VkDisplayPlaneInfo2KHR - Structure defining the intended configuration of a display plane

Description

Note

This parameter also implicitly specifies a display.

planeIndex is the plane which the application intends to use with the display.

The members of VkDisplayPlaneInfo2KHR correspond to the arguments to vkGetDisplayPlaneCapabilitiesKHR, with sType and pNext added for extensibility.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DISPLAY_PLANE_INFO_2_KHR

pNext must be NULL

mode must be a valid VkDisplayModeKHR handle

Host Synchronization

Host access to mode must be externally synchronized

See Also

VkDisplayModeKHR, VkStructureType, vkGetDisplayPlaneCapabilities2KHR

Constructors

DisplayPlaneInfo2KHR

Fields

next :: Maybe SomeVkStruct

mode :: DisplayModeKHR

planeIndex :: Word32

withCStructDisplayPlaneProperties2KHR :: DisplayPlaneProperties2KHR -> (VkDisplayPlaneProperties2KHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkDisplayPlaneProperties2KHR and marshal a DisplayPlaneProperties2KHR into it. The VkDisplayPlaneProperties2KHR is only valid inside the provided computation and must not be returned out of it.

fromCStructDisplayPlaneProperties2KHR :: VkDisplayPlaneProperties2KHR -> IO DisplayPlaneProperties2KHR

A function to read a VkDisplayPlaneProperties2KHR and all additional structures in the pointer chain into a DisplayPlaneProperties2KHR.

data DisplayPlaneProperties2KHR

VkDisplayPlaneProperties2KHR - Structure describing an available display plane

Valid Usage (Implicit)

See Also

VkDisplayPlanePropertiesKHR, VkStructureType, vkGetPhysicalDeviceDisplayPlaneProperties2KHR

Constructors

DisplayPlaneProperties2KHR

Fields

next :: Maybe SomeVkStruct

displayPlaneProperties :: DisplayPlanePropertiesKHR

withCStructDisplayProperties2KHR :: DisplayProperties2KHR -> (VkDisplayProperties2KHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkDisplayProperties2KHR and marshal a DisplayProperties2KHR into it. The VkDisplayProperties2KHR is only valid inside the provided computation and must not be returned out of it.

fromCStructDisplayProperties2KHR :: VkDisplayProperties2KHR -> IO DisplayProperties2KHR

A function to read a VkDisplayProperties2KHR and all additional structures in the pointer chain into a DisplayProperties2KHR.

data DisplayProperties2KHR

VkDisplayProperties2KHR - Structure describing an available display device

Valid Usage (Implicit)

See Also

VkDisplayPropertiesKHR, VkStructureType, vkGetPhysicalDeviceDisplayProperties2KHR

Constructors

DisplayProperties2KHR

Fields

next :: Maybe SomeVkStruct

displayProperties :: DisplayPropertiesKHR

getNumDisplayModeProperties2KHR :: PhysicalDevice -> DisplayKHR -> IO (VkResult, Word32)

vkGetDisplayModeProperties2KHR - Query information about the available display modes.

Parameters

physicalDevice is the physical device associated with display.

display is the display to query.

pPropertyCount is a pointer to an integer related to the number of display modes available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayModeProperties2KHR structures.

Description

vkGetDisplayModeProperties2KHR behaves similarly to vkGetDisplayModePropertiesKHR, with the ability to return extended information via chained output structures.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

display must be a valid VkDisplayKHR handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayModeProperties2KHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayKHR, VkDisplayModeProperties2KHR, VkPhysicalDevice

getDisplayModeProperties2KHR :: PhysicalDevice -> DisplayKHR -> Word32 -> IO (VkResult, Vector DisplayModeProperties2KHR)

vkGetDisplayModeProperties2KHR - Query information about the available display modes.

Parameters

physicalDevice is the physical device associated with display.

display is the display to query.

pPropertyCount is a pointer to an integer related to the number of display modes available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayModeProperties2KHR structures.

Description

vkGetDisplayModeProperties2KHR behaves similarly to vkGetDisplayModePropertiesKHR, with the ability to return extended information via chained output structures.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

display must be a valid VkDisplayKHR handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayModeProperties2KHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayKHR, VkDisplayModeProperties2KHR, VkPhysicalDevice

getAllDisplayModeProperties2KHR :: PhysicalDevice -> DisplayKHR -> IO (Vector DisplayModeProperties2KHR)

Returns all the values available from getDisplayModeProperties2KHR.

getDisplayPlaneCapabilities2KHR :: PhysicalDevice -> DisplayPlaneInfo2KHR -> IO DisplayPlaneCapabilities2KHR

vkGetDisplayPlaneCapabilities2KHR - Query capabilities of a mode and plane combination

Parameters

physicalDevice is the physical device associated with pDisplayPlaneInfo.

pDisplayPlaneInfo is a pointer to an instance of the VkDisplayPlaneInfo2KHR structure describing the plane and mode.

pCapabilities is a pointer to a VkDisplayPlaneCapabilities2KHR structure in which the capabilities are returned.

Description

vkGetDisplayPlaneCapabilities2KHR behaves similarly to vkGetDisplayPlaneCapabilitiesKHR, with the ability to specify extended inputs via chained input structures, and to return extended information via chained output structures.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayPlaneCapabilities2KHR, VkDisplayPlaneInfo2KHR, VkPhysicalDevice

getNumPhysicalDeviceDisplayPlaneProperties2KHR :: PhysicalDevice -> IO (VkResult, Word32)

vkGetPhysicalDeviceDisplayPlaneProperties2KHR - Query information about the available display planes.

Parameters

physicalDevice is a physical device.

pPropertyCount is a pointer to an integer related to the number of display planes available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayPlaneProperties2KHR structures.

Description

vkGetPhysicalDeviceDisplayPlaneProperties2KHR behaves similarly to vkGetPhysicalDeviceDisplayPlanePropertiesKHR, with the ability to return extended information via chained output structures.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayPlaneProperties2KHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayPlaneProperties2KHR, VkPhysicalDevice

getPhysicalDeviceDisplayPlaneProperties2KHR :: PhysicalDevice -> Word32 -> IO (VkResult, Vector DisplayPlaneProperties2KHR)

vkGetPhysicalDeviceDisplayPlaneProperties2KHR - Query information about the available display planes.

Parameters

physicalDevice is a physical device.

pPropertyCount is a pointer to an integer related to the number of display planes available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayPlaneProperties2KHR structures.

Description

vkGetPhysicalDeviceDisplayPlaneProperties2KHR behaves similarly to vkGetPhysicalDeviceDisplayPlanePropertiesKHR, with the ability to return extended information via chained output structures.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayPlaneProperties2KHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayPlaneProperties2KHR, VkPhysicalDevice

getAllPhysicalDeviceDisplayPlaneProperties2KHR :: PhysicalDevice -> IO (Vector DisplayPlaneProperties2KHR)

Returns all the values available from getPhysicalDeviceDisplayPlaneProperties2KHR.

getNumPhysicalDeviceDisplayProperties2KHR :: PhysicalDevice -> IO (VkResult, Word32)

vkGetPhysicalDeviceDisplayProperties2KHR - Query information about the available displays

Parameters

physicalDevice is a physical device.

pPropertyCount is a pointer to an integer related to the number of display devices available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayProperties2KHR structures.

Description

vkGetPhysicalDeviceDisplayProperties2KHR behaves similarly to vkGetPhysicalDeviceDisplayPropertiesKHR, with the ability to return extended information via chained output structures.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayProperties2KHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayProperties2KHR, VkPhysicalDevice

getPhysicalDeviceDisplayProperties2KHR :: PhysicalDevice -> Word32 -> IO (VkResult, Vector DisplayProperties2KHR)

vkGetPhysicalDeviceDisplayProperties2KHR - Query information about the available displays

Parameters

physicalDevice is a physical device.

pPropertyCount is a pointer to an integer related to the number of display devices available or queried, as described below.

pProperties is either NULL or a pointer to an array of VkDisplayProperties2KHR structures.

Description

vkGetPhysicalDeviceDisplayProperties2KHR behaves similarly to vkGetPhysicalDeviceDisplayPropertiesKHR, with the ability to return extended information via chained output structures.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkDisplayProperties2KHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDisplayProperties2KHR, VkPhysicalDevice

getAllPhysicalDeviceDisplayProperties2KHR :: PhysicalDevice -> IO (Vector DisplayProperties2KHR)

Returns all the values available from getPhysicalDeviceDisplayProperties2KHR.

pattern KHR_GET_DISPLAY_PROPERTIES_2_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_GET_DISPLAY_PROPERTIES_2_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_DISPLAY_PROPERTIES_2_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DISPLAY_PLANE_PROPERTIES_2_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DISPLAY_MODE_PROPERTIES_2_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DISPLAY_PLANE_INFO_2_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DISPLAY_PLANE_CAPABILITIES_2_KHR :: a ~ StructureType => a