Graphics.Vulkan.Extensions.VK_KHR_get_surface_capabilities2

Documentation

withCStructPhysicalDeviceSurfaceInfo2KHR :: PhysicalDeviceSurfaceInfo2KHR -> (VkPhysicalDeviceSurfaceInfo2KHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceSurfaceInfo2KHR and marshal a PhysicalDeviceSurfaceInfo2KHR into it. The VkPhysicalDeviceSurfaceInfo2KHR is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceSurfaceInfo2KHR :: VkPhysicalDeviceSurfaceInfo2KHR -> IO PhysicalDeviceSurfaceInfo2KHR

A function to read a VkPhysicalDeviceSurfaceInfo2KHR and all additional structures in the pointer chain into a PhysicalDeviceSurfaceInfo2KHR.

data PhysicalDeviceSurfaceInfo2KHR

VkPhysicalDeviceSurfaceInfo2KHR - Structure specifying a surface and related swapchain creation parameters

Description

The members of VkPhysicalDeviceSurfaceInfo2KHR correspond to the arguments to vkGetPhysicalDeviceSurfaceCapabilitiesKHR, with sType and pNext added for extensibility.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_SURFACE_INFO_2_KHR

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkSurfaceFullScreenExclusiveInfoEXT or VkSurfaceFullScreenExclusiveWin32InfoEXT

Each sType member in the pNext chain must be unique

surface must be a valid VkSurfaceKHR handle

See Also

VkStructureType, VkSurfaceKHR, vkGetDeviceGroupSurfacePresentModes2EXT, vkGetPhysicalDeviceSurfaceCapabilities2KHR, vkGetPhysicalDeviceSurfaceFormats2KHR, vkGetPhysicalDeviceSurfacePresentModes2EXT

Constructors

PhysicalDeviceSurfaceInfo2KHR

Fields

next :: Maybe SomeVkStruct

surface :: SurfaceKHR

withCStructSurfaceCapabilities2KHR :: SurfaceCapabilities2KHR -> (VkSurfaceCapabilities2KHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkSurfaceCapabilities2KHR and marshal a SurfaceCapabilities2KHR into it. The VkSurfaceCapabilities2KHR is only valid inside the provided computation and must not be returned out of it.

fromCStructSurfaceCapabilities2KHR :: VkSurfaceCapabilities2KHR -> IO SurfaceCapabilities2KHR

A function to read a VkSurfaceCapabilities2KHR and all additional structures in the pointer chain into a SurfaceCapabilities2KHR.

data SurfaceCapabilities2KHR

VkSurfaceCapabilities2KHR - Structure describing capabilities of a surface

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_SURFACE_CAPABILITIES_2_KHR

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkDisplayNativeHdrSurfaceCapabilitiesAMD, VkSharedPresentSurfaceCapabilitiesKHR, VkSurfaceCapabilitiesFullScreenExclusiveEXT, or VkSurfaceProtectedCapabilitiesKHR

Each sType member in the pNext chain must be unique

See Also

VkStructureType, VkSurfaceCapabilitiesKHR, vkGetPhysicalDeviceSurfaceCapabilities2KHR

Constructors

SurfaceCapabilities2KHR

Fields

next :: Maybe SomeVkStruct

surfaceCapabilities :: SurfaceCapabilitiesKHR

withCStructSurfaceFormat2KHR :: SurfaceFormat2KHR -> (VkSurfaceFormat2KHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkSurfaceFormat2KHR and marshal a SurfaceFormat2KHR into it. The VkSurfaceFormat2KHR is only valid inside the provided computation and must not be returned out of it.

fromCStructSurfaceFormat2KHR :: VkSurfaceFormat2KHR -> IO SurfaceFormat2KHR

A function to read a VkSurfaceFormat2KHR and all additional structures in the pointer chain into a SurfaceFormat2KHR.

data SurfaceFormat2KHR

VkSurfaceFormat2KHR - Structure describing a supported swapchain format tuple

Valid Usage (Implicit)

See Also

VkStructureType, VkSurfaceFormatKHR, vkGetPhysicalDeviceSurfaceFormats2KHR

Constructors

SurfaceFormat2KHR

Fields

next :: Maybe SomeVkStruct

surfaceFormat :: SurfaceFormatKHR

getPhysicalDeviceSurfaceCapabilities2KHR :: PhysicalDevice -> PhysicalDeviceSurfaceInfo2KHR -> IO SurfaceCapabilities2KHR

vkGetPhysicalDeviceSurfaceCapabilities2KHR - Reports capabilities of a surface on a physical device

Parameters

physicalDevice is the physical device that will be associated with the swapchain to be created, as described for vkCreateSwapchainKHR.

pSurfaceInfo points to an instance of the VkPhysicalDeviceSurfaceInfo2KHR structure, describing the surface and other fixed parameters that would be consumed by vkCreateSwapchainKHR.

pSurfaceCapabilities points to an instance of the VkSurfaceCapabilities2KHR structure in which the capabilities are returned.

Description

vkGetPhysicalDeviceSurfaceCapabilities2KHR behaves similarly to vkGetPhysicalDeviceSurfaceCapabilitiesKHR, with the ability to specify extended inputs via chained input structures, and to return extended information via chained output structures.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

See Also

VkPhysicalDevice, VkPhysicalDeviceSurfaceInfo2KHR, VkSurfaceCapabilities2KHR

getNumPhysicalDeviceSurfaceFormats2KHR :: PhysicalDevice -> PhysicalDeviceSurfaceInfo2KHR -> IO (VkResult, Word32)

vkGetPhysicalDeviceSurfaceFormats2KHR - Query color formats supported by surface

Parameters

physicalDevice is the physical device that will be associated with the swapchain to be created, as described for vkCreateSwapchainKHR.

pSurfaceInfo points to an instance of the VkPhysicalDeviceSurfaceInfo2KHR structure, describing the surface and other fixed parameters that would be consumed by vkCreateSwapchainKHR.

pSurfaceFormatCount is a pointer to an integer related to the number of format tuples available or queried, as described below.

pSurfaceFormats is either NULL or a pointer to an array of VkSurfaceFormat2KHR structures.

Description

If pSurfaceFormats is NULL, then the number of format tuples supported for the given surface is returned in pSurfaceFormatCount. The number of format tuples supported will be greater than or equal to 1. Otherwise, pSurfaceFormatCount must point to a variable set by the user to the number of elements in the pSurfaceFormats array, and on return the variable is overwritten with the number of structures actually written to pSurfaceFormats. If the value of pSurfaceFormatCount is less than the number of format tuples supported, at most pSurfaceFormatCount structures will be written. If pSurfaceFormatCount is smaller than the number of format tuples supported for the surface parameters described in pSurfaceInfo, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pSurfaceInfo must be a valid pointer to a valid VkPhysicalDeviceSurfaceInfo2KHR structure

pSurfaceFormatCount must be a valid pointer to a uint32_t value

If the value referenced by pSurfaceFormatCount is not 0, and pSurfaceFormats is not NULL, pSurfaceFormats must be a valid pointer to an array of pSurfaceFormatCount VkSurfaceFormat2KHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

See Also

VkPhysicalDevice, VkPhysicalDeviceSurfaceInfo2KHR, VkSurfaceFormat2KHR

getPhysicalDeviceSurfaceFormats2KHR :: PhysicalDevice -> PhysicalDeviceSurfaceInfo2KHR -> Word32 -> IO (VkResult, Vector SurfaceFormat2KHR)

vkGetPhysicalDeviceSurfaceFormats2KHR - Query color formats supported by surface

Parameters

physicalDevice is the physical device that will be associated with the swapchain to be created, as described for vkCreateSwapchainKHR.

pSurfaceInfo points to an instance of the VkPhysicalDeviceSurfaceInfo2KHR structure, describing the surface and other fixed parameters that would be consumed by vkCreateSwapchainKHR.

pSurfaceFormatCount is a pointer to an integer related to the number of format tuples available or queried, as described below.

pSurfaceFormats is either NULL or a pointer to an array of VkSurfaceFormat2KHR structures.

Description

If pSurfaceFormats is NULL, then the number of format tuples supported for the given surface is returned in pSurfaceFormatCount. The number of format tuples supported will be greater than or equal to 1. Otherwise, pSurfaceFormatCount must point to a variable set by the user to the number of elements in the pSurfaceFormats array, and on return the variable is overwritten with the number of structures actually written to pSurfaceFormats. If the value of pSurfaceFormatCount is less than the number of format tuples supported, at most pSurfaceFormatCount structures will be written. If pSurfaceFormatCount is smaller than the number of format tuples supported for the surface parameters described in pSurfaceInfo, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pSurfaceInfo must be a valid pointer to a valid VkPhysicalDeviceSurfaceInfo2KHR structure

pSurfaceFormatCount must be a valid pointer to a uint32_t value

If the value referenced by pSurfaceFormatCount is not 0, and pSurfaceFormats is not NULL, pSurfaceFormats must be a valid pointer to an array of pSurfaceFormatCount VkSurfaceFormat2KHR structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

See Also

VkPhysicalDevice, VkPhysicalDeviceSurfaceInfo2KHR, VkSurfaceFormat2KHR

getAllPhysicalDeviceSurfaceFormats2KHR :: PhysicalDevice -> PhysicalDeviceSurfaceInfo2KHR -> IO (Vector SurfaceFormat2KHR)

Returns all the values available from getPhysicalDeviceSurfaceFormats2KHR.

pattern KHR_GET_SURFACE_CAPABILITIES_2_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_GET_SURFACE_CAPABILITIES_2_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SURFACE_INFO_2_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SURFACE_CAPABILITIES_2_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SURFACE_FORMAT_2_KHR :: a ~ StructureType => a