Graphics.Vulkan.Extensions.VK_KHR_image_format_list

Documentation

withCStructImageFormatListCreateInfoKHR :: ImageFormatListCreateInfoKHR -> (VkImageFormatListCreateInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkImageFormatListCreateInfoKHR and marshal a ImageFormatListCreateInfoKHR into it. The VkImageFormatListCreateInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructImageFormatListCreateInfoKHR :: VkImageFormatListCreateInfoKHR -> IO ImageFormatListCreateInfoKHR

A function to read a VkImageFormatListCreateInfoKHR and all additional structures in the pointer chain into a ImageFormatListCreateInfoKHR.

data ImageFormatListCreateInfoKHR

VkImageFormatListCreateInfoKHR - Specify that an image can be used with a particular set of formats

Description

If viewFormatCount is zero, pViewFormats is ignored and the image is created as if the VkImageFormatListCreateInfoKHR structure were not included in the pNext list of VkImageCreateInfo.

Valid Usage

If viewFormatCount is not 0, all of the formats in the pViewFormats array must be compatible with the format specified in the format field of VkImageCreateInfo, as described in the compatibility table.

If VkImageCreateInfo::flags does not contain VK_IMAGE_CREATE_MUTABLE_FORMAT_BIT, viewFormatCount must be 0 or 1.

If viewFormatCount is not 0, VkImageCreateInfo::format must be in pViewFormats.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMAGE_FORMAT_LIST_CREATE_INFO_KHR

If viewFormatCount is not 0, pViewFormats must be a valid pointer to an array of viewFormatCount valid VkFormat values

See Also

VkFormat, VkStructureType

Constructors

ImageFormatListCreateInfoKHR

Fields

next :: Maybe SomeVkStruct

viewFormats :: Vector Format

pattern KHR_IMAGE_FORMAT_LIST_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_IMAGE_FORMAT_LIST_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_IMAGE_FORMAT_LIST_CREATE_INFO_KHR :: a ~ StructureType => a