Graphics.Vulkan.Extensions.VK_KHR_incremental_present

Documentation

withCStructPresentRegionKHR :: PresentRegionKHR -> (VkPresentRegionKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkPresentRegionKHR and marshal a PresentRegionKHR into it. The VkPresentRegionKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructPresentRegionKHR :: VkPresentRegionKHR -> IO PresentRegionKHR

A function to read a VkPresentRegionKHR and all additional structures in the pointer chain into a PresentRegionKHR.

data PresentRegionKHR

VkPresentRegionKHR - Structure containing rectangular region changed by vkQueuePresentKHR for a given VkImage

Valid Usage (Implicit)

If rectangleCount is not 0, and pRectangles is not NULL, pRectangles must be a valid pointer to an array of rectangleCount valid VkRectLayerKHR structures

See Also

VkPresentRegionsKHR, VkRectLayerKHR

Constructors

PresentRegionKHR

Fields

rectangles :: Maybe (Vector RectLayerKHR)

withCStructPresentRegionsKHR :: PresentRegionsKHR -> (VkPresentRegionsKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkPresentRegionsKHR and marshal a PresentRegionsKHR into it. The VkPresentRegionsKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructPresentRegionsKHR :: VkPresentRegionsKHR -> IO PresentRegionsKHR

A function to read a VkPresentRegionsKHR and all additional structures in the pointer chain into a PresentRegionsKHR.

data PresentRegionsKHR

VkPresentRegionsKHR - Structure hint of rectangular regions changed by vkQueuePresentKHR

Valid Usage

swapchainCount must be the same value as VkPresentInfoKHR::swapchainCount, where VkPresentInfoKHR is in the pNext chain of this VkPresentRegionsKHR structure

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PRESENT_REGIONS_KHR

If pRegions is not NULL, pRegions must be a valid pointer to an array of swapchainCount valid VkPresentRegionKHR structures

swapchainCount must be greater than 0

See Also

VkPresentRegionKHR, VkStructureType

Constructors

PresentRegionsKHR

Fields

next :: Maybe SomeVkStruct

regions :: Maybe (Vector PresentRegionKHR)

withCStructRectLayerKHR :: RectLayerKHR -> (VkRectLayerKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkRectLayerKHR and marshal a RectLayerKHR into it. The VkRectLayerKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructRectLayerKHR :: VkRectLayerKHR -> IO RectLayerKHR

A function to read a VkRectLayerKHR and all additional structures in the pointer chain into a RectLayerKHR.

data RectLayerKHR

VkRectLayerKHR - Structure containing a rectangle, including layer, changed by vkQueuePresentKHR for a given VkImage

Valid Usage

The sum of offset and extent must be no greater than the imageExtent member of the VkSwapchainCreateInfoKHR structure given to vkCreateSwapchainKHR.

layer must be less than imageArrayLayers member of the VkSwapchainCreateInfoKHR structure given to vkCreateSwapchainKHR.

Some platforms allow the size of a surface to change, and then scale the pixels of the image to fit the surface. VkRectLayerKHR specifies pixels of the swapchain’s image(s), which will be constant for the life of the swapchain.

See Also

VkExtent2D, VkOffset2D, VkPresentRegionKHR

Constructors

RectLayerKHR

Fields

offset :: Offset2D

extent :: Extent2D

layer :: Word32

pattern KHR_INCREMENTAL_PRESENT_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_INCREMENTAL_PRESENT_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PRESENT_REGIONS_KHR :: a ~ StructureType => a