Graphics.Vulkan.Extensions.VK_KHR_maintenance2

Documentation

type ImageViewUsageCreateInfoKHR = ImageViewUsageCreateInfo

type InputAttachmentAspectReferenceKHR = InputAttachmentAspectReference

type PhysicalDevicePointClippingPropertiesKHR = PhysicalDevicePointClippingProperties

type PipelineTessellationDomainOriginStateCreateInfoKHR = PipelineTessellationDomainOriginStateCreateInfo

type RenderPassInputAttachmentAspectCreateInfoKHR = RenderPassInputAttachmentAspectCreateInfo

pattern IMAGE_CREATE_BLOCK_TEXEL_VIEW_COMPATIBLE_BIT_KHR :: VkImageCreateFlagBits

pattern IMAGE_CREATE_EXTENDED_USAGE_BIT_KHR :: VkImageCreateFlagBits

pattern IMAGE_LAYOUT_DEPTH_ATTACHMENT_STENCIL_READ_ONLY_OPTIMAL_KHR :: VkImageLayout

pattern IMAGE_LAYOUT_DEPTH_READ_ONLY_STENCIL_ATTACHMENT_OPTIMAL_KHR :: VkImageLayout

pattern KHR_MAINTENANCE2_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_MAINTENANCE2_SPEC_VERSION :: Integral a => a

pattern POINT_CLIPPING_BEHAVIOR_ALL_CLIP_PLANES_KHR :: VkPointClippingBehavior

pattern POINT_CLIPPING_BEHAVIOR_USER_CLIP_PLANES_ONLY_KHR :: VkPointClippingBehavior

pattern STRUCTURE_TYPE_IMAGE_VIEW_USAGE_CREATE_INFO_KHR :: VkStructureType

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_POINT_CLIPPING_PROPERTIES_KHR :: VkStructureType

pattern STRUCTURE_TYPE_PIPELINE_TESSELLATION_DOMAIN_ORIGIN_STATE_CREATE_INFO_KHR :: VkStructureType

pattern STRUCTURE_TYPE_RENDER_PASS_INPUT_ATTACHMENT_ASPECT_CREATE_INFO_KHR :: VkStructureType

pattern TESSELLATION_DOMAIN_ORIGIN_LOWER_LEFT_KHR :: VkTessellationDomainOrigin

pattern TESSELLATION_DOMAIN_ORIGIN_UPPER_LEFT_KHR :: VkTessellationDomainOrigin

type PointClippingBehaviorKHR = PointClippingBehavior

type TessellationDomainOriginKHR = TessellationDomainOrigin

pattern IMAGE_CREATE_BLOCK_TEXEL_VIEW_COMPATIBLE_BIT :: a ~ ImageCreateFlagBits => a

pattern IMAGE_CREATE_EXTENDED_USAGE_BIT :: a ~ ImageCreateFlagBits => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_POINT_CLIPPING_PROPERTIES :: a ~ StructureType => a

pattern STRUCTURE_TYPE_RENDER_PASS_INPUT_ATTACHMENT_ASPECT_CREATE_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_IMAGE_VIEW_USAGE_CREATE_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PIPELINE_TESSELLATION_DOMAIN_ORIGIN_STATE_CREATE_INFO :: a ~ StructureType => a

pattern IMAGE_LAYOUT_DEPTH_READ_ONLY_STENCIL_ATTACHMENT_OPTIMAL :: a ~ ImageLayout => a

pattern IMAGE_LAYOUT_DEPTH_ATTACHMENT_STENCIL_READ_ONLY_OPTIMAL :: a ~ ImageLayout => a

pattern POINT_CLIPPING_BEHAVIOR_ALL_CLIP_PLANES :: a ~ PointClippingBehavior => a

VK_POINT_CLIPPING_BEHAVIOR_ALL_CLIP_PLANES specifies that the primitive is discarded if the vertex lies outside any clip plane, including the planes bounding the view volume.

pattern POINT_CLIPPING_BEHAVIOR_USER_CLIP_PLANES_ONLY :: a ~ PointClippingBehavior => a

VK_POINT_CLIPPING_BEHAVIOR_USER_CLIP_PLANES_ONLY specifies that the primitive is discarded only if the vertex lies outside any user clip plane.

pattern TESSELLATION_DOMAIN_ORIGIN_UPPER_LEFT :: a ~ TessellationDomainOrigin => a

VK_TESSELLATION_DOMAIN_ORIGIN_UPPER_LEFT specifies that the origin of the domain space is in the upper left corner, as shown in figure https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#img-tessellation-topology-ul.

pattern TESSELLATION_DOMAIN_ORIGIN_LOWER_LEFT :: a ~ TessellationDomainOrigin => a

VK_TESSELLATION_DOMAIN_ORIGIN_LOWER_LEFT specifies that the origin of the domain space is in the lower left corner, as shown in figure https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#img-tessellation-topology-ll.