Graphics.Vulkan.Extensions.VK_KHR_performance_query

Documentation

enumeratePhysicalDeviceQueueFamilyPerformanceQueryCountersKHR :: PhysicalDevice -> ("queueFamilyIndex" ::: Word32) -> IO (Result, "counters" ::: Vector PerformanceCounterKHR, "counterDescriptions" ::: Vector PerformanceCounterDescriptionKHR)

vkEnumeratePhysicalDeviceQueueFamilyPerformanceQueryCountersKHR - Reports properties of the performance query counters available on a queue family of a device

Parameters

PhysicalDevice is the handle to the physical device whose queue family performance query counter properties will be queried.

queueFamilyIndex is the index into the queue family of the physical device we want to get properties for.

pCounterCount is a pointer to an integer related to the number of counters available or queried, as described below.

pCounters is either NULL or a pointer to an array of PerformanceCounterKHR structures.

pCounterDescriptions is either NULL or a pointer to an array of PerformanceCounterDescriptionKHR structures.

Description

If pCounters is NULL and pCounterDescriptions is NULL, then the number of counters available is returned in pCounterCount. Otherwise, pCounterCount must point to a variable set by the user to the number of elements in the pCounters, pCounterDescriptions, or both arrays and on return the variable is overwritten with the number of structures actually written out. If pCounterCount is less than the number of counters available, at most pCounterCount structures will be written and INCOMPLETE will be returned instead of SUCCESS.

Valid Usage (Implicit)

PhysicalDevice must be a valid PhysicalDevice handle

pCounterCount must be a valid pointer to a uint32_t value

If the value referenced by pCounterCount is not 0, and pCounters is not NULL, pCounters must be a valid pointer to an array of pCounterCount PerformanceCounterKHR structures

If the value referenced by pCounterCount is not 0, and pCounterDescriptions is not NULL, pCounterDescriptions must be a valid pointer to an array of pCounterCount PerformanceCounterDescriptionKHR structures

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_INITIALIZATION_FAILED

See Also

PerformanceCounterDescriptionKHR, PerformanceCounterKHR, PhysicalDevice

getPhysicalDeviceQueueFamilyPerformanceQueryPassesKHR :: PhysicalDevice -> ("performanceQueryCreateInfo" ::: QueryPoolPerformanceCreateInfoKHR) -> IO ("numPasses" ::: Word32)

vkGetPhysicalDeviceQueueFamilyPerformanceQueryPassesKHR - Reports the number of passes require for a performance query pool type

Parameters

PhysicalDevice is the handle to the physical device whose queue family performance query counter properties will be queried.

pPerformanceQueryCreateInfo is a pointer to a QueryPoolPerformanceCreateInfoKHR of the performance query that is to be created.

pNumPasses is a pointer to an integer related to the number of passes required to query the performance query pool, as described below.

Description

The pPerformanceQueryCreateInfo member QueryPoolPerformanceCreateInfoKHR::queueFamilyIndex must be a queue family of PhysicalDevice. The number of passes required to capture the counters specified in the pPerformanceQueryCreateInfo member QueryPoolPerformanceCreateInfoKHR::pCounters is returned in pNumPasses.

Valid Usage (Implicit)

See Also

PhysicalDevice, QueryPoolPerformanceCreateInfoKHR

acquireProfilingLockKHR :: Device -> AcquireProfilingLockInfoKHR -> IO ()

vkAcquireProfilingLockKHR - Acquires the profiling lock

Parameters

Device is the logical device to profile.

pInfo is a pointer to a AcquireProfilingLockInfoKHR structure which contains information about how the profiling is to be acquired.

Description

Implementations may allow multiple actors to hold the profiling lock concurrently.

Return Codes

Success

SUCCESS

Failure

TIMEOUT

See Also

AcquireProfilingLockInfoKHR, Device

releaseProfilingLockKHR :: Device -> IO ()

vkReleaseProfilingLockKHR - Releases the profiling lock

Parameters

Device is the logical device to cease profiling on.

Valid Usage

The profiling lock of Device must have been held via a previous successful call to acquireProfilingLockKHR

Valid Usage (Implicit)

Device must be a valid Device handle

See Also

Device

data PhysicalDevicePerformanceQueryFeaturesKHR

VkPhysicalDevicePerformanceQueryFeaturesKHR - Structure describing performance query support for an implementation

Members

The members of the PhysicalDevicePerformanceQueryFeaturesKHR structure describe the following implementation-dependent features:

Valid Usage (Implicit)

To query supported performance counter query pool features, call getPhysicalDeviceFeatures2 with a PhysicalDevicePerformanceQueryFeaturesKHR structure included in the pNext chain of its pFeatures parameter. The PhysicalDevicePerformanceQueryFeaturesKHR structure can also be included in the pNext chain of a DeviceCreateInfo structure, in which case it controls which additional features are enabled in the device.

See Also

Bool32, StructureType

Constructors

PhysicalDevicePerformanceQueryFeaturesKHR

Fields

performanceCounterQueryPools :: Bool

performanceCounterQueryPools is TRUE if the physical device supports performance counter query pools.

performanceCounterMultipleQueryPools :: Bool

performanceCounterMultipleQueryPools is 'Graphics.Vulkan.Core10.BaseType.TRUE'\` if the physical device supports using multiple performance query pools in a primary command buffer and secondary command buffers executed within it.

data PhysicalDevicePerformanceQueryPropertiesKHR

VkPhysicalDevicePerformanceQueryPropertiesKHR - Structure describing performance query properties for an implementation

Members

The members of the PhysicalDevicePerformanceQueryPropertiesKHR structure describe the following implementation-dependent properties:

Valid Usage (Implicit)

If the PhysicalDevicePerformanceQueryPropertiesKHR structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent properties.

See Also

Bool32, StructureType

Constructors

PhysicalDevicePerformanceQueryPropertiesKHR

Fields

allowCommandBufferQueryCopies :: Bool

allowCommandBufferQueryCopies is TRUE if the performance query pools are allowed to be used with cmdCopyQueryPoolResults.

data PerformanceCounterKHR

VkPerformanceCounterKHR - Structure providing information about a counter

Valid Usage (Implicit)

See Also

PerformanceCounterScopeKHR, PerformanceCounterStorageKHR, PerformanceCounterUnitKHR, StructureType, enumeratePhysicalDeviceQueueFamilyPerformanceQueryCountersKHR

Constructors

PerformanceCounterKHR

Fields

unit :: PerformanceCounterUnitKHR

unit is a PerformanceCounterUnitKHR specifying the unit that the counter data will record.

scope :: PerformanceCounterScopeKHR

scope is a PerformanceCounterScopeKHR specifying the scope that the counter belongs to.

storage :: PerformanceCounterStorageKHR

storage is a PerformanceCounterStorageKHR specifying the storage type that the counter’s data uses.

uuid :: ByteString

uuid is an array of size UUID_SIZE, containing 8-bit values that represent a universally unique identifier for the counter of the physical device.

data PerformanceCounterDescriptionKHR

VkPerformanceCounterDescriptionKHR - Structure providing more detailed information about a counter

Valid Usage (Implicit)

See Also

PerformanceCounterDescriptionFlagsKHR, StructureType, enumeratePhysicalDeviceQueueFamilyPerformanceQueryCountersKHR

Constructors

PerformanceCounterDescriptionKHR

Fields

flags :: PerformanceCounterDescriptionFlagsKHR

Flags is a bitmask of PerformanceCounterDescriptionFlagBitsKHR indicating the usage behavior for the counter.

name :: ByteString

name is an array of size MAX_DESCRIPTION_SIZE, containing a null-terminated UTF-8 string specifying the name of the counter.

category :: ByteString

category is an array of size MAX_DESCRIPTION_SIZE, containing a null-terminated UTF-8 string specifying the category of the counter.

description :: ByteString

description is an array of size MAX_DESCRIPTION_SIZE, containing a null-terminated UTF-8 string specifying the description of the counter.

data QueryPoolPerformanceCreateInfoKHR

VkQueryPoolPerformanceCreateInfoKHR - Structure specifying parameters of a newly created performance query pool

Valid Usage

queueFamilyIndex must be a valid queue family index of the device

The performanceCounterQueryPools feature must be enabled

Each element of pCounterIndices must be in the range of counters reported by enumeratePhysicalDeviceQueueFamilyPerformanceQueryCountersKHR for the queue family specified in queueFamilyIndex

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_QUERY_POOL_PERFORMANCE_CREATE_INFO_KHR

pCounterIndices must be a valid pointer to an array of counterIndexCount uint32_t values

counterIndexCount must be greater than 0

See Also

StructureType, getPhysicalDeviceQueueFamilyPerformanceQueryPassesKHR

Constructors

QueryPoolPerformanceCreateInfoKHR

Fields

queueFamilyIndex :: Word32

queueFamilyIndex is the queue family index to create this performance query pool for.

counterIndices :: Vector Word32

pCounterIndices is the array of indices into the enumeratePhysicalDeviceQueueFamilyPerformanceQueryCountersKHR::pCounters to enable in this performance query pool.

data AcquireProfilingLockInfoKHR

VkAcquireProfilingLockInfoKHR - Structure specifying parameters to acquire the profiling lock

Valid Usage (Implicit)

If timeout is 0, acquireProfilingLockKHR will not block while attempting to acquire the profling lock. If timeout is UINT64_MAX, the function will not return until the profiling lock was acquired.

See Also

AcquireProfilingLockFlagsKHR, StructureType, acquireProfilingLockKHR

Constructors

AcquireProfilingLockInfoKHR

Fields

flags :: AcquireProfilingLockFlagsKHR

Flags must be 0

timeout :: Word64

timeout indicates how long the function waits, in nanoseconds, if the profiling lock is not available.

data PerformanceQuerySubmitInfoKHR

VkPerformanceQuerySubmitInfoKHR - Structure indicating which counter pass index is active for performance queries

Description

If the SubmitInfo::pNext chain does not include this structure, the batch defaults to use counter pass index 0.

Valid Usage (Implicit)

See Also

StructureType

Constructors

PerformanceQuerySubmitInfoKHR

Fields

counterPassIndex :: Word32

counterPassIndex must be less than the number of counter passes required by any queries within the batch. The required number of counter passes for a performance query is obtained by calling getPhysicalDeviceQueueFamilyPerformanceQueryPassesKHR

data PerformanceCounterResultKHR

Constructors

CounterInt32 Int32

CounterInt64 Int64

CounterUint32 Word32

CounterUint64 Word64

CounterFloat32 Float

CounterFloat64 Double

newtype PerformanceCounterScopeKHR

VkPerformanceCounterScopeKHR - Supported counter scope types

See Also

PerformanceCounterKHR

Constructors

PerformanceCounterScopeKHR Int32

Bundled Patterns

pattern PERFORMANCE_COUNTER_SCOPE_COMMAND_BUFFER_KHR :: PerformanceCounterScopeKHR
PERFORMANCE_COUNTER_SCOPE_COMMAND_BUFFER_KHR - the performance counter scope is a single complete command buffer.

pattern PERFORMANCE_COUNTER_SCOPE_RENDER_PASS_KHR :: PerformanceCounterScopeKHR
PERFORMANCE_COUNTER_SCOPE_RENDER_PASS_KHR - the performance counter scope is zero or more complete render passes. The performance query containing the performance counter must begin and end outside a render pass instance.

pattern PERFORMANCE_COUNTER_SCOPE_COMMAND_KHR :: PerformanceCounterScopeKHR
PERFORMANCE_COUNTER_SCOPE_COMMAND_KHR - the performance counter scope is zero or more commands.

newtype PerformanceCounterUnitKHR

VkPerformanceCounterUnitKHR - Supported counter unit types

See Also

PerformanceCounterKHR

Constructors

PerformanceCounterUnitKHR Int32

Bundled Patterns

pattern PERFORMANCE_COUNTER_UNIT_GENERIC_KHR :: PerformanceCounterUnitKHR
PERFORMANCE_COUNTER_UNIT_GENERIC_KHR - the performance counter unit is a generic data point.

pattern PERFORMANCE_COUNTER_UNIT_PERCENTAGE_KHR :: PerformanceCounterUnitKHR
PERFORMANCE_COUNTER_UNIT_PERCENTAGE_KHR - the performance counter unit is a percentage (%).

pattern PERFORMANCE_COUNTER_UNIT_NANOSECONDS_KHR :: PerformanceCounterUnitKHR
PERFORMANCE_COUNTER_UNIT_NANOSECONDS_KHR - the performance counter unit is a value of nanoseconds (ns).

pattern PERFORMANCE_COUNTER_UNIT_BYTES_KHR :: PerformanceCounterUnitKHR
PERFORMANCE_COUNTER_UNIT_BYTES_KHR - the performance counter unit is a value of bytes.

pattern PERFORMANCE_COUNTER_UNIT_BYTES_PER_SECOND_KHR :: PerformanceCounterUnitKHR
PERFORMANCE_COUNTER_UNIT_BYTES_PER_SECOND_KHR - the performance counter unit is a value of bytes/s.

pattern PERFORMANCE_COUNTER_UNIT_KELVIN_KHR :: PerformanceCounterUnitKHR
PERFORMANCE_COUNTER_UNIT_KELVIN_KHR - the performance counter unit is a temperature reported in Kelvin.

pattern PERFORMANCE_COUNTER_UNIT_WATTS_KHR :: PerformanceCounterUnitKHR
PERFORMANCE_COUNTER_UNIT_WATTS_KHR - the performance counter unit is a value of watts (W).

pattern PERFORMANCE_COUNTER_UNIT_VOLTS_KHR :: PerformanceCounterUnitKHR
PERFORMANCE_COUNTER_UNIT_VOLTS_KHR - the performance counter unit is a value of volts (V).

pattern PERFORMANCE_COUNTER_UNIT_AMPS_KHR :: PerformanceCounterUnitKHR
PERFORMANCE_COUNTER_UNIT_AMPS_KHR - the performance counter unit is a value of amps (A).

pattern PERFORMANCE_COUNTER_UNIT_HERTZ_KHR :: PerformanceCounterUnitKHR
PERFORMANCE_COUNTER_UNIT_HERTZ_KHR - the performance counter unit is a value of hertz (Hz).

pattern PERFORMANCE_COUNTER_UNIT_CYCLES_KHR :: PerformanceCounterUnitKHR
PERFORMANCE_COUNTER_UNIT_CYCLES_KHR - the performance counter unit is a value of cycles.

newtype PerformanceCounterStorageKHR

VkPerformanceCounterStorageKHR - Supported counter storage types

See Also

PerformanceCounterKHR

Constructors

PerformanceCounterStorageKHR Int32

Bundled Patterns

pattern PERFORMANCE_COUNTER_STORAGE_INT32_KHR :: PerformanceCounterStorageKHR
PERFORMANCE_COUNTER_STORAGE_INT32_KHR - the performance counter storage is a 32-bit signed integer.

pattern PERFORMANCE_COUNTER_STORAGE_INT64_KHR :: PerformanceCounterStorageKHR
PERFORMANCE_COUNTER_STORAGE_INT64_KHR - the performance counter storage is a 64-bit signed integer.

pattern PERFORMANCE_COUNTER_STORAGE_UINT32_KHR :: PerformanceCounterStorageKHR
PERFORMANCE_COUNTER_STORAGE_UINT32_KHR - the performance counter storage is a 32-bit unsigned integer.

pattern PERFORMANCE_COUNTER_STORAGE_UINT64_KHR :: PerformanceCounterStorageKHR
PERFORMANCE_COUNTER_STORAGE_UINT64_KHR - the performance counter storage is a 64-bit unsigned integer.

pattern PERFORMANCE_COUNTER_STORAGE_FLOAT32_KHR :: PerformanceCounterStorageKHR
PERFORMANCE_COUNTER_STORAGE_FLOAT32_KHR - the performance counter storage is a 32-bit floating-point.

pattern PERFORMANCE_COUNTER_STORAGE_FLOAT64_KHR :: PerformanceCounterStorageKHR
PERFORMANCE_COUNTER_STORAGE_FLOAT64_KHR - the performance counter storage is a 64-bit floating-point.

newtype PerformanceCounterDescriptionFlagBitsKHR

VkPerformanceCounterDescriptionFlagBitsKHR - Bitmask specifying usage behavior for a counter

See Also

PerformanceCounterDescriptionFlagsKHR

Constructors

PerformanceCounterDescriptionFlagBitsKHR Flags

Bundled Patterns

pattern PERFORMANCE_COUNTER_DESCRIPTION_PERFORMANCE_IMPACTING_KHR :: PerformanceCounterDescriptionFlagBitsKHR
PERFORMANCE_COUNTER_DESCRIPTION_PERFORMANCE_IMPACTING_KHR specifies that recording the counter may have a noticeable performance impact.

pattern PERFORMANCE_COUNTER_DESCRIPTION_CONCURRENTLY_IMPACTED_KHR :: PerformanceCounterDescriptionFlagBitsKHR
PERFORMANCE_COUNTER_DESCRIPTION_CONCURRENTLY_IMPACTED_KHR specifies that concurrently recording the counter while other submitted command buffers are running may impact the accuracy of the recording.

type PerformanceCounterDescriptionFlagsKHR = PerformanceCounterDescriptionFlagBitsKHR

newtype AcquireProfilingLockFlagBitsKHR

VkAcquireProfilingLockFlagBitsKHR - Reserved for future use

See Also

AcquireProfilingLockFlagsKHR

Constructors

AcquireProfilingLockFlagBitsKHR Flags

type AcquireProfilingLockFlagsKHR = AcquireProfilingLockFlagBitsKHR

type KHR_PERFORMANCE_QUERY_SPEC_VERSION = 1

pattern KHR_PERFORMANCE_QUERY_SPEC_VERSION :: forall a. Integral a => a

type KHR_PERFORMANCE_QUERY_EXTENSION_NAME = "VK_KHR_performance_query"

pattern KHR_PERFORMANCE_QUERY_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a