Graphics.Vulkan.Extensions.VK_KHR_sampler_ycbcr_conversion

Documentation

type BindImagePlaneMemoryInfoKHR = BindImagePlaneMemoryInfo

type ImagePlaneMemoryRequirementsInfoKHR = ImagePlaneMemoryRequirementsInfo

type PhysicalDeviceSamplerYcbcrConversionFeaturesKHR = PhysicalDeviceSamplerYcbcrConversionFeatures

type SamplerYcbcrConversionCreateInfoKHR = SamplerYcbcrConversionCreateInfo

type SamplerYcbcrConversionImageFormatPropertiesKHR = SamplerYcbcrConversionImageFormatProperties

type SamplerYcbcrConversionInfoKHR = SamplerYcbcrConversionInfo

type SamplerYcbcrConversionKHR = SamplerYcbcrConversion

createSamplerYcbcrConversionKHR :: Device -> SamplerYcbcrConversionCreateInfo -> Maybe AllocationCallbacks -> IO SamplerYcbcrConversion

destroySamplerYcbcrConversionKHR :: Device -> SamplerYcbcrConversion -> Maybe AllocationCallbacks -> IO ()

pattern CHROMA_LOCATION_COSITED_EVEN_KHR :: VkChromaLocation

pattern CHROMA_LOCATION_MIDPOINT_KHR :: VkChromaLocation

pattern DEBUG_REPORT_OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION_KHR_EXT :: VkDebugReportObjectTypeEXT

pattern FORMAT_B10X6G10X6R10X6G10X6_422_UNORM_4PACK16_KHR :: VkFormat

pattern FORMAT_B12X4G12X4R12X4G12X4_422_UNORM_4PACK16_KHR :: VkFormat

pattern FORMAT_B16G16R16G16_422_UNORM_KHR :: VkFormat

pattern FORMAT_B8G8R8G8_422_UNORM_KHR :: VkFormat

pattern FORMAT_FEATURE_COSITED_CHROMA_SAMPLES_BIT_KHR :: VkFormatFeatureFlagBits

pattern FORMAT_FEATURE_DISJOINT_BIT_KHR :: VkFormatFeatureFlagBits

pattern FORMAT_FEATURE_MIDPOINT_CHROMA_SAMPLES_BIT_KHR :: VkFormatFeatureFlagBits

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_BIT_KHR :: VkFormatFeatureFlagBits

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_FORCEABLE_BIT_KHR :: VkFormatFeatureFlagBits

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_LINEAR_FILTER_BIT_KHR :: VkFormatFeatureFlagBits

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_SEPARATE_RECONSTRUCTION_FILTER_BIT_KHR :: VkFormatFeatureFlagBits

pattern FORMAT_G10X6B10X6G10X6R10X6_422_UNORM_4PACK16_KHR :: VkFormat

pattern FORMAT_G10X6_B10X6R10X6_2PLANE_420_UNORM_3PACK16_KHR :: VkFormat

pattern FORMAT_G10X6_B10X6R10X6_2PLANE_422_UNORM_3PACK16_KHR :: VkFormat

pattern FORMAT_G10X6_B10X6_R10X6_3PLANE_420_UNORM_3PACK16_KHR :: VkFormat

pattern FORMAT_G10X6_B10X6_R10X6_3PLANE_422_UNORM_3PACK16_KHR :: VkFormat

pattern FORMAT_G10X6_B10X6_R10X6_3PLANE_444_UNORM_3PACK16_KHR :: VkFormat

pattern FORMAT_G12X4B12X4G12X4R12X4_422_UNORM_4PACK16_KHR :: VkFormat

pattern FORMAT_G12X4_B12X4R12X4_2PLANE_420_UNORM_3PACK16_KHR :: VkFormat

pattern FORMAT_G12X4_B12X4R12X4_2PLANE_422_UNORM_3PACK16_KHR :: VkFormat

pattern FORMAT_G12X4_B12X4_R12X4_3PLANE_420_UNORM_3PACK16_KHR :: VkFormat

pattern FORMAT_G12X4_B12X4_R12X4_3PLANE_422_UNORM_3PACK16_KHR :: VkFormat

pattern FORMAT_G12X4_B12X4_R12X4_3PLANE_444_UNORM_3PACK16_KHR :: VkFormat

pattern FORMAT_G16B16G16R16_422_UNORM_KHR :: VkFormat

pattern FORMAT_G16_B16R16_2PLANE_420_UNORM_KHR :: VkFormat

pattern FORMAT_G16_B16R16_2PLANE_422_UNORM_KHR :: VkFormat

pattern FORMAT_G16_B16_R16_3PLANE_420_UNORM_KHR :: VkFormat

pattern FORMAT_G16_B16_R16_3PLANE_422_UNORM_KHR :: VkFormat

pattern FORMAT_G16_B16_R16_3PLANE_444_UNORM_KHR :: VkFormat

pattern FORMAT_G8B8G8R8_422_UNORM_KHR :: VkFormat

pattern FORMAT_G8_B8R8_2PLANE_420_UNORM_KHR :: VkFormat

pattern FORMAT_G8_B8R8_2PLANE_422_UNORM_KHR :: VkFormat

pattern FORMAT_G8_B8_R8_3PLANE_420_UNORM_KHR :: VkFormat

pattern FORMAT_G8_B8_R8_3PLANE_422_UNORM_KHR :: VkFormat

pattern FORMAT_G8_B8_R8_3PLANE_444_UNORM_KHR :: VkFormat

pattern FORMAT_R10X6G10X6B10X6A10X6_UNORM_4PACK16_KHR :: VkFormat

pattern FORMAT_R10X6G10X6_UNORM_2PACK16_KHR :: VkFormat

pattern FORMAT_R10X6_UNORM_PACK16_KHR :: VkFormat

pattern FORMAT_R12X4G12X4B12X4A12X4_UNORM_4PACK16_KHR :: VkFormat

pattern FORMAT_R12X4G12X4_UNORM_2PACK16_KHR :: VkFormat

pattern FORMAT_R12X4_UNORM_PACK16_KHR :: VkFormat

pattern IMAGE_ASPECT_PLANE_0_BIT_KHR :: VkImageAspectFlagBits

pattern IMAGE_ASPECT_PLANE_1_BIT_KHR :: VkImageAspectFlagBits

pattern IMAGE_ASPECT_PLANE_2_BIT_KHR :: VkImageAspectFlagBits

pattern IMAGE_CREATE_DISJOINT_BIT_KHR :: VkImageCreateFlagBits

pattern KHR_SAMPLER_YCBCR_CONVERSION_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_SAMPLER_YCBCR_CONVERSION_SPEC_VERSION :: Integral a => a

pattern OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION_KHR :: VkObjectType

pattern SAMPLER_YCBCR_MODEL_CONVERSION_RGB_IDENTITY_KHR :: VkSamplerYcbcrModelConversion

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_2020_KHR :: VkSamplerYcbcrModelConversion

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_601_KHR :: VkSamplerYcbcrModelConversion

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_709_KHR :: VkSamplerYcbcrModelConversion

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_IDENTITY_KHR :: VkSamplerYcbcrModelConversion

pattern SAMPLER_YCBCR_RANGE_ITU_FULL_KHR :: VkSamplerYcbcrRange

pattern SAMPLER_YCBCR_RANGE_ITU_NARROW_KHR :: VkSamplerYcbcrRange

pattern STRUCTURE_TYPE_BIND_IMAGE_PLANE_MEMORY_INFO_KHR :: VkStructureType

pattern STRUCTURE_TYPE_IMAGE_PLANE_MEMORY_REQUIREMENTS_INFO_KHR :: VkStructureType

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SAMPLER_YCBCR_CONVERSION_FEATURES_KHR :: VkStructureType

pattern STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_CREATE_INFO_KHR :: VkStructureType

pattern STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_IMAGE_FORMAT_PROPERTIES_KHR :: VkStructureType

pattern STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_INFO_KHR :: VkStructureType

type SamplerYcbcrModelConversionKHR = SamplerYcbcrModelConversion

type SamplerYcbcrRangeKHR = SamplerYcbcrRange

type ChromaLocationKHR = ChromaLocation

pattern STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_CREATE_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_BIND_IMAGE_PLANE_MEMORY_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_IMAGE_PLANE_MEMORY_REQUIREMENTS_INFO :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SAMPLER_YCBCR_CONVERSION_FEATURES :: a ~ StructureType => a

pattern STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_IMAGE_FORMAT_PROPERTIES :: a ~ StructureType => a

pattern DEBUG_REPORT_OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION_EXT :: a ~ DebugReportObjectTypeEXT => a

pattern OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION :: a ~ ObjectType => a

pattern FORMAT_G8B8G8R8_422_UNORM :: a ~ Format => a

pattern FORMAT_B8G8R8G8_422_UNORM :: a ~ Format => a

pattern FORMAT_G8_B8_R8_3PLANE_420_UNORM :: a ~ Format => a

pattern FORMAT_G8_B8R8_2PLANE_420_UNORM :: a ~ Format => a

pattern FORMAT_G8_B8_R8_3PLANE_422_UNORM :: a ~ Format => a

pattern FORMAT_G8_B8R8_2PLANE_422_UNORM :: a ~ Format => a

pattern FORMAT_G8_B8_R8_3PLANE_444_UNORM :: a ~ Format => a

pattern FORMAT_R10X6_UNORM_PACK16 :: a ~ Format => a

pattern FORMAT_R10X6G10X6_UNORM_2PACK16 :: a ~ Format => a

pattern FORMAT_R10X6G10X6B10X6A10X6_UNORM_4PACK16 :: a ~ Format => a

pattern FORMAT_G10X6B10X6G10X6R10X6_422_UNORM_4PACK16 :: a ~ Format => a

pattern FORMAT_B10X6G10X6R10X6G10X6_422_UNORM_4PACK16 :: a ~ Format => a

pattern FORMAT_G10X6_B10X6_R10X6_3PLANE_420_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G10X6_B10X6R10X6_2PLANE_420_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G10X6_B10X6_R10X6_3PLANE_422_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G10X6_B10X6R10X6_2PLANE_422_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G10X6_B10X6_R10X6_3PLANE_444_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_R12X4_UNORM_PACK16 :: a ~ Format => a

pattern FORMAT_R12X4G12X4_UNORM_2PACK16 :: a ~ Format => a

pattern FORMAT_R12X4G12X4B12X4A12X4_UNORM_4PACK16 :: a ~ Format => a

pattern FORMAT_G12X4B12X4G12X4R12X4_422_UNORM_4PACK16 :: a ~ Format => a

pattern FORMAT_B12X4G12X4R12X4G12X4_422_UNORM_4PACK16 :: a ~ Format => a

pattern FORMAT_G12X4_B12X4_R12X4_3PLANE_420_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G12X4_B12X4R12X4_2PLANE_420_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G12X4_B12X4_R12X4_3PLANE_422_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G12X4_B12X4R12X4_2PLANE_422_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G12X4_B12X4_R12X4_3PLANE_444_UNORM_3PACK16 :: a ~ Format => a

pattern FORMAT_G16B16G16R16_422_UNORM :: a ~ Format => a

pattern FORMAT_B16G16R16G16_422_UNORM :: a ~ Format => a

pattern FORMAT_G16_B16_R16_3PLANE_420_UNORM :: a ~ Format => a

pattern FORMAT_G16_B16R16_2PLANE_420_UNORM :: a ~ Format => a

pattern FORMAT_G16_B16_R16_3PLANE_422_UNORM :: a ~ Format => a

pattern FORMAT_G16_B16R16_2PLANE_422_UNORM :: a ~ Format => a

pattern FORMAT_G16_B16_R16_3PLANE_444_UNORM :: a ~ Format => a

pattern IMAGE_ASPECT_PLANE_0_BIT :: a ~ ImageAspectFlagBits => a

pattern IMAGE_ASPECT_PLANE_1_BIT :: a ~ ImageAspectFlagBits => a

pattern IMAGE_ASPECT_PLANE_2_BIT :: a ~ ImageAspectFlagBits => a

pattern IMAGE_CREATE_DISJOINT_BIT :: a ~ ImageCreateFlagBits => a

pattern FORMAT_FEATURE_MIDPOINT_CHROMA_SAMPLES_BIT :: a ~ FormatFeatureFlagBits => a

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_LINEAR_FILTER_BIT :: a ~ FormatFeatureFlagBits => a

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_SEPARATE_RECONSTRUCTION_FILTER_BIT :: a ~ FormatFeatureFlagBits => a

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_BIT :: a ~ FormatFeatureFlagBits => a

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_FORCEABLE_BIT :: a ~ FormatFeatureFlagBits => a

pattern FORMAT_FEATURE_DISJOINT_BIT :: a ~ FormatFeatureFlagBits => a

pattern FORMAT_FEATURE_COSITED_CHROMA_SAMPLES_BIT :: a ~ FormatFeatureFlagBits => a

pattern SAMPLER_YCBCR_MODEL_CONVERSION_RGB_IDENTITY :: a ~ SamplerYcbcrModelConversion => a

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_IDENTITY :: a ~ SamplerYcbcrModelConversion => a

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_709 :: a ~ SamplerYcbcrModelConversion => a

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_601 :: a ~ SamplerYcbcrModelConversion => a

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_2020 :: a ~ SamplerYcbcrModelConversion => a

pattern SAMPLER_YCBCR_RANGE_ITU_FULL :: a ~ SamplerYcbcrRange => a

VK_SAMPLER_YCBCR_RANGE_ITU_FULL specifies that the full range of the encoded values are valid and interpreted according to the ITU “full range” quantization rules.

pattern SAMPLER_YCBCR_RANGE_ITU_NARROW :: a ~ SamplerYcbcrRange => a

VK_SAMPLER_YCBCR_RANGE_ITU_NARROW specifies that headroom and foot room are reserved in the numerical range of encoded values, and the remaining values are expanded according to the ITU “narrow range” quantization rules.

pattern CHROMA_LOCATION_COSITED_EVEN :: a ~ ChromaLocation => a

VK_CHROMA_LOCATION_COSITED_EVEN specifies that downsampled chroma samples are aligned with luma samples with even coordinates.

pattern CHROMA_LOCATION_MIDPOINT :: a ~ ChromaLocation => a

VK_CHROMA_LOCATION_MIDPOINT specifies that downsampled chroma samples are located half way between each even luma sample and the nearest higher odd luma sample.