Graphics.Vulkan.Extensions.VK_KHR_shader_atomic_int64

Documentation

withCStructPhysicalDeviceShaderAtomicInt64FeaturesKHR :: PhysicalDeviceShaderAtomicInt64FeaturesKHR -> (VkPhysicalDeviceShaderAtomicInt64FeaturesKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceShaderAtomicInt64FeaturesKHR and marshal a PhysicalDeviceShaderAtomicInt64FeaturesKHR into it. The VkPhysicalDeviceShaderAtomicInt64FeaturesKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceShaderAtomicInt64FeaturesKHR :: VkPhysicalDeviceShaderAtomicInt64FeaturesKHR -> IO PhysicalDeviceShaderAtomicInt64FeaturesKHR

A function to read a VkPhysicalDeviceShaderAtomicInt64FeaturesKHR and all additional structures in the pointer chain into a PhysicalDeviceShaderAtomicInt64FeaturesKHR.

data PhysicalDeviceShaderAtomicInt64FeaturesKHR

VkPhysicalDeviceShaderAtomicInt64FeaturesKHR - Structure describing features supported by VK_KHR_shader_atomic_int64

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceShaderAtomicInt64FeaturesKHR

Fields

next :: Maybe SomeVkStruct

shaderBufferInt64Atomics :: Bool

shaderSharedInt64Atomics :: Bool

pattern KHR_SHADER_ATOMIC_INT64_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_SHADER_ATOMIC_INT64_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_ATOMIC_INT64_FEATURES_KHR :: a ~ StructureType => a