Graphics.Vulkan.Extensions.VK_KHR_shader_float16_int8

Documentation

withCStructPhysicalDeviceFloat16Int8FeaturesKHR :: PhysicalDeviceFloat16Int8FeaturesKHR -> (VkPhysicalDeviceFloat16Int8FeaturesKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceFloat16Int8FeaturesKHR and marshal a PhysicalDeviceFloat16Int8FeaturesKHR into it. The VkPhysicalDeviceFloat16Int8FeaturesKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceFloat16Int8FeaturesKHR :: VkPhysicalDeviceFloat16Int8FeaturesKHR -> IO PhysicalDeviceFloat16Int8FeaturesKHR

A function to read a VkPhysicalDeviceFloat16Int8FeaturesKHR and all additional structures in the pointer chain into a PhysicalDeviceFloat16Int8FeaturesKHR.

data PhysicalDeviceFloat16Int8FeaturesKHR

VkPhysicalDeviceFloat16Int8FeaturesKHR - Structure describing features supported by VK_KHR_shader_float16_int8

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceFloat16Int8FeaturesKHR

Fields

next :: Maybe SomeVkStruct

shaderFloat16 :: Bool

shaderInt8 :: Bool

pattern KHR_SHADER_FLOAT16_INT8_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_SHADER_FLOAT16_INT8_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_FLOAT16_INT8_FEATURES_KHR :: a ~ StructureType => a