Graphics.Vulkan.Extensions.VK_KHR_shader_float_controls

Documentation

withCStructPhysicalDeviceFloatControlsPropertiesKHR :: PhysicalDeviceFloatControlsPropertiesKHR -> (VkPhysicalDeviceFloatControlsPropertiesKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceFloatControlsPropertiesKHR and marshal a PhysicalDeviceFloatControlsPropertiesKHR into it. The VkPhysicalDeviceFloatControlsPropertiesKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceFloatControlsPropertiesKHR :: VkPhysicalDeviceFloatControlsPropertiesKHR -> IO PhysicalDeviceFloatControlsPropertiesKHR

A function to read a VkPhysicalDeviceFloatControlsPropertiesKHR and all additional structures in the pointer chain into a PhysicalDeviceFloatControlsPropertiesKHR.

data PhysicalDeviceFloatControlsPropertiesKHR

VkPhysicalDeviceFloatControlsPropertiesKHR - Structure describing properties supported by VK_KHR_shader_float_controls

Description

If the VkPhysicalDeviceFloatControlsPropertiesKHR structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceFloatControlsPropertiesKHR

Fields

next :: Maybe SomeVkStruct

separateDenormSettings :: Bool

separateRoundingModeSettings :: Bool

shaderSignedZeroInfNanPreserveFloat16 :: Bool

shaderSignedZeroInfNanPreserveFloat32 :: Bool

shaderSignedZeroInfNanPreserveFloat64 :: Bool

shaderDenormPreserveFloat16 :: Bool

shaderDenormPreserveFloat32 :: Bool

shaderDenormPreserveFloat64 :: Bool

shaderDenormFlushToZeroFloat16 :: Bool

shaderDenormFlushToZeroFloat32 :: Bool

shaderDenormFlushToZeroFloat64 :: Bool

shaderRoundingModeRTEFloat16 :: Bool

shaderRoundingModeRTEFloat32 :: Bool

shaderRoundingModeRTEFloat64 :: Bool

shaderRoundingModeRTZFloat16 :: Bool

shaderRoundingModeRTZFloat32 :: Bool

shaderRoundingModeRTZFloat64 :: Bool

pattern KHR_SHADER_FLOAT_CONTROLS_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_SHADER_FLOAT_CONTROLS_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_FLOAT_CONTROLS_PROPERTIES_KHR :: a ~ StructureType => a