Graphics.Vulkan.Extensions.VK_KHR_shared_presentable_image

Documentation

withCStructSharedPresentSurfaceCapabilitiesKHR :: SharedPresentSurfaceCapabilitiesKHR -> (VkSharedPresentSurfaceCapabilitiesKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkSharedPresentSurfaceCapabilitiesKHR and marshal a SharedPresentSurfaceCapabilitiesKHR into it. The VkSharedPresentSurfaceCapabilitiesKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructSharedPresentSurfaceCapabilitiesKHR :: VkSharedPresentSurfaceCapabilitiesKHR -> IO SharedPresentSurfaceCapabilitiesKHR

A function to read a VkSharedPresentSurfaceCapabilitiesKHR and all additional structures in the pointer chain into a SharedPresentSurfaceCapabilitiesKHR.

data SharedPresentSurfaceCapabilitiesKHR

VkSharedPresentSurfaceCapabilitiesKHR - structure describing capabilities of a surface for shared presentation

Valid Usage (Implicit)

See Also

VkImageUsageFlags, VkStructureType

Constructors

SharedPresentSurfaceCapabilitiesKHR

Fields

next :: Maybe SomeVkStruct

sharedPresentSupportedUsageFlags :: ImageUsageFlags

getSwapchainStatusKHR :: Device -> SwapchainKHR -> IO VkResult

vkGetSwapchainStatusKHR - Get a swapchain’s status

Parameters

device is the device associated with swapchain.

swapchain is the swapchain to query.

Valid Usage (Implicit)

device must be a valid VkDevice handle

swapchain must be a valid VkSwapchainKHR handle

Both of device, and swapchain must have been created, allocated, or retrieved from the same VkInstance

Host Synchronization

Host access to swapchain must be externally synchronized

Return Codes

Success

- VK_SUCCESS

VK_SUBOPTIMAL_KHR

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_DEVICE_LOST

VK_ERROR_OUT_OF_DATE_KHR

VK_ERROR_SURFACE_LOST_KHR

VK_ERROR_FULL_SCREEN_EXCLUSIVE_MODE_LOST_EXT

See Also

VkDevice, VkSwapchainKHR

pattern KHR_SHARED_PRESENTABLE_IMAGE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_SHARED_PRESENTABLE_IMAGE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_SHARED_PRESENT_SURFACE_CAPABILITIES_KHR :: a ~ StructureType => a

pattern PRESENT_MODE_SHARED_DEMAND_REFRESH_KHR :: a ~ PresentModeKHR => a

pattern PRESENT_MODE_SHARED_CONTINUOUS_REFRESH_KHR :: a ~ PresentModeKHR => a

pattern IMAGE_LAYOUT_SHARED_PRESENT_KHR :: a ~ ImageLayout => a