Graphics.Vulkan.Extensions.VK_KHR_surface

Documentation

type ColorSpaceKHR = VkColorSpaceKHR

VkColorSpaceKHR - supported color space of the presentation engine

See Also

VkSurfaceFormatKHR, VkSwapchainCreateInfoKHR

pattern COLOR_SPACE_SRGB_NONLINEAR_KHR :: a ~ ColorSpaceKHR => a

VK_COLOR_SPACE_SRGB_NONLINEAR_KHR specifies support for the sRGB color space.

pattern COLOR_SPACE_DISPLAY_P3_NONLINEAR_EXT :: a ~ ColorSpaceKHR => a

pattern COLOR_SPACE_EXTENDED_SRGB_LINEAR_EXT :: a ~ ColorSpaceKHR => a

pattern COLOR_SPACE_DCI_P3_LINEAR_EXT :: a ~ ColorSpaceKHR => a

pattern COLOR_SPACE_DCI_P3_NONLINEAR_EXT :: a ~ ColorSpaceKHR => a

pattern COLOR_SPACE_BT709_LINEAR_EXT :: a ~ ColorSpaceKHR => a

pattern COLOR_SPACE_BT709_NONLINEAR_EXT :: a ~ ColorSpaceKHR => a

pattern COLOR_SPACE_BT2020_LINEAR_EXT :: a ~ ColorSpaceKHR => a

pattern COLOR_SPACE_HDR10_ST2084_EXT :: a ~ ColorSpaceKHR => a

pattern COLOR_SPACE_DOLBYVISION_EXT :: a ~ ColorSpaceKHR => a

pattern COLOR_SPACE_HDR10_HLG_EXT :: a ~ ColorSpaceKHR => a

pattern COLOR_SPACE_ADOBERGB_LINEAR_EXT :: a ~ ColorSpaceKHR => a

pattern COLOR_SPACE_ADOBERGB_NONLINEAR_EXT :: a ~ ColorSpaceKHR => a

pattern COLOR_SPACE_PASS_THROUGH_EXT :: a ~ ColorSpaceKHR => a

pattern COLOR_SPACE_EXTENDED_SRGB_NONLINEAR_EXT :: a ~ ColorSpaceKHR => a

pattern COLOR_SPACE_DISPLAY_NATIVE_AMD :: a ~ ColorSpaceKHR => a

type CompositeAlphaFlagBitsKHR = VkCompositeAlphaFlagBitsKHR

VkCompositeAlphaFlagBitsKHR - alpha compositing modes supported on a device

Description

These values are described as follows:

See Also

VkCompositeAlphaFlagsKHR, VkSwapchainCreateInfoKHR

pattern COMPOSITE_ALPHA_OPAQUE_BIT_KHR :: a ~ CompositeAlphaFlagBitsKHR => a

VK_COMPOSITE_ALPHA_OPAQUE_BIT_KHR: The alpha channel, if it exists, of the images is ignored in the compositing process. Instead, the image is treated as if it has a constant alpha of 1.0.

pattern COMPOSITE_ALPHA_PRE_MULTIPLIED_BIT_KHR :: a ~ CompositeAlphaFlagBitsKHR => a

VK_COMPOSITE_ALPHA_PRE_MULTIPLIED_BIT_KHR: The alpha channel, if it exists, of the images is respected in the compositing process. The non-alpha channels of the image are expected to already be multiplied by the alpha channel by the application.

pattern COMPOSITE_ALPHA_POST_MULTIPLIED_BIT_KHR :: a ~ CompositeAlphaFlagBitsKHR => a

VK_COMPOSITE_ALPHA_POST_MULTIPLIED_BIT_KHR: The alpha channel, if it exists, of the images is respected in the compositing process. The non-alpha channels of the image are not expected to already be multiplied by the alpha channel by the application; instead, the compositor will multiply the non-alpha channels of the image by the alpha channel during compositing.

pattern COMPOSITE_ALPHA_INHERIT_BIT_KHR :: a ~ CompositeAlphaFlagBitsKHR => a

VK_COMPOSITE_ALPHA_INHERIT_BIT_KHR: The way in which the presentation engine treats the alpha channel in the images is unknown to the Vulkan API. Instead, the application is responsible for setting the composite alpha blending mode using native window system commands. If the application does not set the blending mode using native window system commands, then a platform-specific default will be used.

type CompositeAlphaFlagsKHR = CompositeAlphaFlagBitsKHR

VkCompositeAlphaFlagsKHR - Bitmask of VkCompositeAlphaFlagBitsKHR

Description

VkCompositeAlphaFlagsKHR is a bitmask type for setting a mask of zero or more VkCompositeAlphaFlagBitsKHR.

See Also

VkCompositeAlphaFlagBitsKHR, VkSurfaceCapabilities2EXT, VkSurfaceCapabilitiesKHR

type PresentModeKHR = VkPresentModeKHR

VkPresentModeKHR - presentation mode supported for a surface

Description

Note

For reference, the mode indicated by VK_PRESENT_MODE_FIFO_KHR is equivalent to the behavior of {wgl|glX|egl}SwapBuffers with a swap interval of 1, while the mode indicated by VK_PRESENT_MODE_FIFO_RELAXED_KHR is equivalent to the behavior of {wgl|glX}SwapBuffers with a swap interval of -1 (from the {WGL|GLX}_EXT_swap_control_tear extensions).

See Also

VkSwapchainCreateInfoKHR, vkGetPhysicalDeviceSurfacePresentModes2EXT, vkGetPhysicalDeviceSurfacePresentModesKHR

pattern PRESENT_MODE_IMMEDIATE_KHR :: a ~ PresentModeKHR => a

VK_PRESENT_MODE_IMMEDIATE_KHR specifies that the presentation engine does not wait for a vertical blanking period to update the current image, meaning this mode may result in visible tearing. No internal queuing of presentation requests is needed, as the requests are applied immediately.

pattern PRESENT_MODE_MAILBOX_KHR :: a ~ PresentModeKHR => a

VK_PRESENT_MODE_MAILBOX_KHR specifies that the presentation engine waits for the next vertical blanking period to update the current image. Tearing cannot be observed. An internal single-entry queue is used to hold pending presentation requests. If the queue is full when a new presentation request is received, the new request replaces the existing entry, and any images associated with the prior entry become available for re-use by the application. One request is removed from the queue and processed during each vertical blanking period in which the queue is non-empty.

pattern PRESENT_MODE_FIFO_KHR :: a ~ PresentModeKHR => a

VK_PRESENT_MODE_FIFO_KHR specifies that the presentation engine waits for the next vertical blanking period to update the current image. Tearing cannot be observed. An internal queue is used to hold pending presentation requests. New requests are appended to the end of the queue, and one request is removed from the beginning of the queue and processed during each vertical blanking period in which the queue is non-empty. This is the only value of presentMode that is required to be supported.

pattern PRESENT_MODE_FIFO_RELAXED_KHR :: a ~ PresentModeKHR => a

VK_PRESENT_MODE_FIFO_RELAXED_KHR specifies that the presentation engine generally waits for the next vertical blanking period to update the current image. If a vertical blanking period has already passed since the last update of the current image then the presentation engine does not wait for another vertical blanking period for the update, meaning this mode may result in visible tearing in this case. This mode is useful for reducing visual stutter with an application that will mostly present a new image before the next vertical blanking period, but may occasionally be late, and present a new image just after the next vertical blanking period. An internal queue is used to hold pending presentation requests. New requests are appended to the end of the queue, and one request is removed from the beginning of the queue and processed during or after each vertical blanking period in which the queue is non-empty.

pattern PRESENT_MODE_SHARED_DEMAND_REFRESH_KHR :: a ~ PresentModeKHR => a

pattern PRESENT_MODE_SHARED_CONTINUOUS_REFRESH_KHR :: a ~ PresentModeKHR => a

withCStructSurfaceCapabilitiesKHR :: SurfaceCapabilitiesKHR -> (VkSurfaceCapabilitiesKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkSurfaceCapabilitiesKHR and marshal a SurfaceCapabilitiesKHR into it. The VkSurfaceCapabilitiesKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructSurfaceCapabilitiesKHR :: VkSurfaceCapabilitiesKHR -> IO SurfaceCapabilitiesKHR

A function to read a VkSurfaceCapabilitiesKHR and all additional structures in the pointer chain into a SurfaceCapabilitiesKHR.

data SurfaceCapabilitiesKHR

VkSurfaceCapabilitiesKHR - Structure describing capabilities of a surface

Description

Note

Formulas such as min(N, maxImageCount) are not correct, since maxImageCount may be zero.

See Also

VkCompositeAlphaFlagsKHR, VkExtent2D, VkImageUsageFlags, VkSurfaceCapabilities2KHR, VkSurfaceTransformFlagBitsKHR, VkSurfaceTransformFlagsKHR, vkGetPhysicalDeviceSurfaceCapabilitiesKHR

Constructors

SurfaceCapabilitiesKHR

Fields

minImageCount :: Word32

maxImageCount :: Word32

currentExtent :: Extent2D

minImageExtent :: Extent2D

maxImageExtent :: Extent2D

maxImageArrayLayers :: Word32

supportedTransforms :: SurfaceTransformFlagsKHR

currentTransform :: SurfaceTransformFlagBitsKHR

supportedCompositeAlpha :: CompositeAlphaFlagsKHR

supportedUsageFlags :: ImageUsageFlags

withCStructSurfaceFormatKHR :: SurfaceFormatKHR -> (VkSurfaceFormatKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkSurfaceFormatKHR and marshal a SurfaceFormatKHR into it. The VkSurfaceFormatKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructSurfaceFormatKHR :: VkSurfaceFormatKHR -> IO SurfaceFormatKHR

A function to read a VkSurfaceFormatKHR and all additional structures in the pointer chain into a SurfaceFormatKHR.

data SurfaceFormatKHR

VkSurfaceFormatKHR - Structure describing a supported swapchain format-color space pair

See Also

VkColorSpaceKHR, VkFormat, VkSurfaceFormat2KHR, vkGetPhysicalDeviceSurfaceFormatsKHR

Constructors

SurfaceFormatKHR

Fields

format :: Format

colorSpace :: ColorSpaceKHR

type SurfaceKHR = VkSurfaceKHR

VkSurfaceKHR - Opaque handle to a surface object

Description

The VK_KHR_surface extension declares the VkSurfaceKHR object, and provides a function for destroying VkSurfaceKHR objects. Separate platform-specific extensions each provide a function for creating a VkSurfaceKHR object for the respective platform. From the application’s perspective this is an opaque handle, just like the handles of other Vulkan objects.

See Also

VkPhysicalDeviceSurfaceInfo2KHR, VkSwapchainCreateInfoKHR, vkCreateAndroidSurfaceKHR, vkCreateDisplayPlaneSurfaceKHR, vkCreateHeadlessSurfaceEXT, vkCreateIOSSurfaceMVK, vkCreateImagePipeSurfaceFUCHSIA, vkCreateMacOSSurfaceMVK, vkCreateMetalSurfaceEXT, vkCreateStreamDescriptorSurfaceGGP, vkCreateViSurfaceNN, vkCreateWaylandSurfaceKHR, vkCreateWin32SurfaceKHR, vkCreateXcbSurfaceKHR, vkCreateXlibSurfaceKHR, vkDestroySurfaceKHR, vkGetDeviceGroupSurfacePresentModesKHR, vkGetPhysicalDevicePresentRectanglesKHR, vkGetPhysicalDeviceSurfaceCapabilities2EXT, vkGetPhysicalDeviceSurfaceCapabilitiesKHR, vkGetPhysicalDeviceSurfaceFormatsKHR, vkGetPhysicalDeviceSurfacePresentModesKHR, vkGetPhysicalDeviceSurfaceSupportKHR

type SurfaceTransformFlagBitsKHR = VkSurfaceTransformFlagBitsKHR

VkSurfaceTransformFlagBitsKHR - presentation transforms supported on a device

See Also

VkDisplaySurfaceCreateInfoKHR, VkSurfaceCapabilities2EXT, VkSurfaceCapabilitiesKHR, VkSurfaceTransformFlagsKHR, VkSwapchainCreateInfoKHR

pattern SURFACE_TRANSFORM_IDENTITY_BIT_KHR :: a ~ SurfaceTransformFlagBitsKHR => a

VK_SURFACE_TRANSFORM_IDENTITY_BIT_KHR specifies that image content is presented without being transformed.

pattern SURFACE_TRANSFORM_ROTATE_90_BIT_KHR :: a ~ SurfaceTransformFlagBitsKHR => a

VK_SURFACE_TRANSFORM_ROTATE_90_BIT_KHR specifies that image content is rotated 90 degrees clockwise.

pattern SURFACE_TRANSFORM_ROTATE_180_BIT_KHR :: a ~ SurfaceTransformFlagBitsKHR => a

VK_SURFACE_TRANSFORM_ROTATE_180_BIT_KHR specifies that image content is rotated 180 degrees clockwise.

pattern SURFACE_TRANSFORM_ROTATE_270_BIT_KHR :: a ~ SurfaceTransformFlagBitsKHR => a

VK_SURFACE_TRANSFORM_ROTATE_270_BIT_KHR specifies that image content is rotated 270 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_BIT_KHR :: a ~ SurfaceTransformFlagBitsKHR => a

VK_SURFACE_TRANSFORM_HORIZONTAL_MIRROR_BIT_KHR specifies that image content is mirrored horizontally.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_90_BIT_KHR :: a ~ SurfaceTransformFlagBitsKHR => a

VK_SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_90_BIT_KHR specifies that image content is mirrored horizontally, then rotated 90 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_180_BIT_KHR :: a ~ SurfaceTransformFlagBitsKHR => a

VK_SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_180_BIT_KHR specifies that image content is mirrored horizontally, then rotated 180 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_270_BIT_KHR :: a ~ SurfaceTransformFlagBitsKHR => a

VK_SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_270_BIT_KHR specifies that image content is mirrored horizontally, then rotated 270 degrees clockwise.

pattern SURFACE_TRANSFORM_INHERIT_BIT_KHR :: a ~ SurfaceTransformFlagBitsKHR => a

VK_SURFACE_TRANSFORM_INHERIT_BIT_KHR specifies that the presentation transform is not specified, and is instead determined by platform-specific considerations and mechanisms outside Vulkan.

type SurfaceTransformFlagsKHR = SurfaceTransformFlagBitsKHR

VkSurfaceTransformFlagsKHR - Bitmask of VkSurfaceTransformFlagBitsKHR

Description

VkSurfaceTransformFlagsKHR is a bitmask type for setting a mask of zero or more VkSurfaceTransformFlagBitsKHR.

See Also

VkDisplayPropertiesKHR, VkSurfaceCapabilities2EXT, VkSurfaceCapabilitiesKHR, VkSurfaceTransformFlagBitsKHR

destroySurfaceKHR :: Instance -> SurfaceKHR -> Maybe AllocationCallbacks -> IO ()

vkDestroySurfaceKHR - Destroy a VkSurfaceKHR object

Parameters

instance is the instance used to create the surface.

surface is the surface to destroy.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

Description

Destroying a VkSurfaceKHR merely severs the connection between Vulkan and the native surface, and does not imply destroying the native surface, closing a window, or similar behavior.

Valid Usage

All VkSwapchainKHR objects created for surface must have been destroyed prior to destroying surface

If VkAllocationCallbacks were provided when surface was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when surface was created, pAllocator must be NULL

Valid Usage (Implicit)

instance must be a valid VkInstance handle

If surface is not VK_NULL_HANDLE, surface must be a valid VkSurfaceKHR handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

If surface is a valid handle, it must have been created, allocated, or retrieved from instance

Host Synchronization

Host access to surface must be externally synchronized

See Also

VkAllocationCallbacks, VkInstance, VkSurfaceKHR

getPhysicalDeviceSurfaceCapabilitiesKHR :: PhysicalDevice -> SurfaceKHR -> IO SurfaceCapabilitiesKHR

vkGetPhysicalDeviceSurfaceCapabilitiesKHR - Query surface capabilities

Parameters

physicalDevice is the physical device that will be associated with the swapchain to be created, as described for vkCreateSwapchainKHR.

surface is the surface that will be associated with the swapchain.

pSurfaceCapabilities is a pointer to an instance of the VkSurfaceCapabilitiesKHR structure in which the capabilities are returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

surface must be a valid VkSurfaceKHR handle

pSurfaceCapabilities must be a valid pointer to a VkSurfaceCapabilitiesKHR structure

Both of physicalDevice, and surface must have been created, allocated, or retrieved from the same VkInstance

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

See Also

VkPhysicalDevice, VkSurfaceCapabilitiesKHR, VkSurfaceKHR

getNumPhysicalDeviceSurfaceFormatsKHR :: PhysicalDevice -> SurfaceKHR -> IO (VkResult, Word32)

vkGetPhysicalDeviceSurfaceFormatsKHR - Query color formats supported by surface

Parameters

physicalDevice is the physical device that will be associated with the swapchain to be created, as described for vkCreateSwapchainKHR.

surface is the surface that will be associated with the swapchain.

pSurfaceFormatCount is a pointer to an integer related to the number of format pairs available or queried, as described below.

pSurfaceFormats is either NULL or a pointer to an array of VkSurfaceFormatKHR structures.

Description

If pSurfaceFormats is NULL, then the number of format pairs supported for the given surface is returned in pSurfaceFormatCount. The number of format pairs supported will be greater than or equal to 1. Otherwise, pSurfaceFormatCount must point to a variable set by the user to the number of elements in the pSurfaceFormats array, and on return the variable is overwritten with the number of structures actually written to pSurfaceFormats. If the value of pSurfaceFormatCount is less than the number of format pairs supported, at most pSurfaceFormatCount structures will be written. If pSurfaceFormatCount is smaller than the number of format pairs supported for the given surface, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

surface must be a valid VkSurfaceKHR handle

pSurfaceFormatCount must be a valid pointer to a uint32_t value

If the value referenced by pSurfaceFormatCount is not 0, and pSurfaceFormats is not NULL, pSurfaceFormats must be a valid pointer to an array of pSurfaceFormatCount VkSurfaceFormatKHR structures

Both of physicalDevice, and surface must have been created, allocated, or retrieved from the same VkInstance

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

See Also

VkPhysicalDevice, VkSurfaceFormatKHR, VkSurfaceKHR

getPhysicalDeviceSurfaceFormatsKHR :: PhysicalDevice -> SurfaceKHR -> Word32 -> IO (VkResult, Vector SurfaceFormatKHR)

vkGetPhysicalDeviceSurfaceFormatsKHR - Query color formats supported by surface

Parameters

physicalDevice is the physical device that will be associated with the swapchain to be created, as described for vkCreateSwapchainKHR.

surface is the surface that will be associated with the swapchain.

pSurfaceFormatCount is a pointer to an integer related to the number of format pairs available or queried, as described below.

pSurfaceFormats is either NULL or a pointer to an array of VkSurfaceFormatKHR structures.

Description

If pSurfaceFormats is NULL, then the number of format pairs supported for the given surface is returned in pSurfaceFormatCount. The number of format pairs supported will be greater than or equal to 1. Otherwise, pSurfaceFormatCount must point to a variable set by the user to the number of elements in the pSurfaceFormats array, and on return the variable is overwritten with the number of structures actually written to pSurfaceFormats. If the value of pSurfaceFormatCount is less than the number of format pairs supported, at most pSurfaceFormatCount structures will be written. If pSurfaceFormatCount is smaller than the number of format pairs supported for the given surface, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

surface must be a valid VkSurfaceKHR handle

pSurfaceFormatCount must be a valid pointer to a uint32_t value

If the value referenced by pSurfaceFormatCount is not 0, and pSurfaceFormats is not NULL, pSurfaceFormats must be a valid pointer to an array of pSurfaceFormatCount VkSurfaceFormatKHR structures

Both of physicalDevice, and surface must have been created, allocated, or retrieved from the same VkInstance

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

See Also

VkPhysicalDevice, VkSurfaceFormatKHR, VkSurfaceKHR

getAllPhysicalDeviceSurfaceFormatsKHR :: PhysicalDevice -> SurfaceKHR -> IO (Vector SurfaceFormatKHR)

Returns all the values available from getPhysicalDeviceSurfaceFormatsKHR.

getNumPhysicalDeviceSurfacePresentModesKHR :: PhysicalDevice -> SurfaceKHR -> IO (VkResult, Word32)

vkGetPhysicalDeviceSurfacePresentModesKHR - Query supported presentation modes

Parameters

physicalDevice is the physical device that will be associated with the swapchain to be created, as described for vkCreateSwapchainKHR.

surface is the surface that will be associated with the swapchain.

pPresentModeCount is a pointer to an integer related to the number of presentation modes available or queried, as described below.

pPresentModes is either NULL or a pointer to an array of VkPresentModeKHR values, indicating the supported presentation modes.

Description

If pPresentModes is NULL, then the number of presentation modes supported for the given surface is returned in pPresentModeCount. Otherwise, pPresentModeCount must point to a variable set by the user to the number of elements in the pPresentModes array, and on return the variable is overwritten with the number of values actually written to pPresentModes. If the value of pPresentModeCount is less than the number of presentation modes supported, at most pPresentModeCount values will be written. If pPresentModeCount is smaller than the number of presentation modes supported for the given surface, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

surface must be a valid VkSurfaceKHR handle

pPresentModeCount must be a valid pointer to a uint32_t value

If the value referenced by pPresentModeCount is not 0, and pPresentModes is not NULL, pPresentModes must be a valid pointer to an array of pPresentModeCount VkPresentModeKHR values

Both of physicalDevice, and surface must have been created, allocated, or retrieved from the same VkInstance

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

See Also

VkPhysicalDevice, VkPresentModeKHR, VkSurfaceKHR

getPhysicalDeviceSurfacePresentModesKHR :: PhysicalDevice -> SurfaceKHR -> Word32 -> IO (VkResult, Vector PresentModeKHR)

vkGetPhysicalDeviceSurfacePresentModesKHR - Query supported presentation modes

Parameters

physicalDevice is the physical device that will be associated with the swapchain to be created, as described for vkCreateSwapchainKHR.

surface is the surface that will be associated with the swapchain.

pPresentModeCount is a pointer to an integer related to the number of presentation modes available or queried, as described below.

pPresentModes is either NULL or a pointer to an array of VkPresentModeKHR values, indicating the supported presentation modes.

Description

If pPresentModes is NULL, then the number of presentation modes supported for the given surface is returned in pPresentModeCount. Otherwise, pPresentModeCount must point to a variable set by the user to the number of elements in the pPresentModes array, and on return the variable is overwritten with the number of values actually written to pPresentModes. If the value of pPresentModeCount is less than the number of presentation modes supported, at most pPresentModeCount values will be written. If pPresentModeCount is smaller than the number of presentation modes supported for the given surface, VK_INCOMPLETE will be returned instead of VK_SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

surface must be a valid VkSurfaceKHR handle

pPresentModeCount must be a valid pointer to a uint32_t value

If the value referenced by pPresentModeCount is not 0, and pPresentModes is not NULL, pPresentModes must be a valid pointer to an array of pPresentModeCount VkPresentModeKHR values

Both of physicalDevice, and surface must have been created, allocated, or retrieved from the same VkInstance

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

See Also

VkPhysicalDevice, VkPresentModeKHR, VkSurfaceKHR

getAllPhysicalDeviceSurfacePresentModesKHR :: PhysicalDevice -> SurfaceKHR -> IO (Vector PresentModeKHR)

Returns all the values available from getPhysicalDeviceSurfacePresentModesKHR.

getPhysicalDeviceSurfaceSupportKHR :: PhysicalDevice -> Word32 -> SurfaceKHR -> IO Bool

vkGetPhysicalDeviceSurfaceSupportKHR - Query if presentation is supported

Parameters

physicalDevice is the physical device.

queueFamilyIndex is the queue family.

surface is the surface.

pSupported is a pointer to a VkBool32, which is set to VK_TRUE to indicate support, and VK_FALSE otherwise.

Valid Usage

queueFamilyIndex must be less than pQueueFamilyPropertyCount returned by vkGetPhysicalDeviceQueueFamilyProperties for the given physicalDevice

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

surface must be a valid VkSurfaceKHR handle

pSupported must be a valid pointer to a VkBool32 value

Both of physicalDevice, and surface must have been created, allocated, or retrieved from the same VkInstance

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

See Also

VkBool32, VkPhysicalDevice, VkSurfaceKHR

pattern KHR_SURFACE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_SURFACE_SPEC_VERSION :: Integral a => a

pattern ERROR_SURFACE_LOST_KHR :: a ~ Result => a

pattern ERROR_NATIVE_WINDOW_IN_USE_KHR :: a ~ Result => a

pattern OBJECT_TYPE_SURFACE_KHR :: a ~ ObjectType => a