Graphics.Vulkan.Extensions.VK_KHR_wayland_surface

Documentation

type WaylandSurfaceCreateFlagsKHR = VkWaylandSurfaceCreateFlagsKHR

withCStructWaylandSurfaceCreateInfoKHR :: WaylandSurfaceCreateInfoKHR -> (VkWaylandSurfaceCreateInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkWaylandSurfaceCreateInfoKHR and marshal a WaylandSurfaceCreateInfoKHR into it. The VkWaylandSurfaceCreateInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructWaylandSurfaceCreateInfoKHR :: VkWaylandSurfaceCreateInfoKHR -> IO WaylandSurfaceCreateInfoKHR

A function to read a VkWaylandSurfaceCreateInfoKHR and all additional structures in the pointer chain into a WaylandSurfaceCreateInfoKHR.

data WaylandSurfaceCreateInfoKHR

VkWaylandSurfaceCreateInfoKHR - Structure specifying parameters of a newly created Wayland surface object

Valid Usage (Implicit)

See Also

VkStructureType, VkWaylandSurfaceCreateFlagsKHR, vkCreateWaylandSurfaceKHR

Constructors

WaylandSurfaceCreateInfoKHR

Fields

next :: Maybe SomeVkStruct

flags :: WaylandSurfaceCreateFlagsKHR

display :: Ptr Wl_display

surface :: Ptr Wl_surface

createWaylandSurfaceKHR :: Instance -> WaylandSurfaceCreateInfoKHR -> Maybe AllocationCallbacks -> IO SurfaceKHR

vkCreateWaylandSurfaceKHR - Create a VkSurfaceKHR object for a Wayland window

Parameters

instance is the instance to associate the surface with.

pCreateInfo is a pointer to an instance of the VkWaylandSurfaceCreateInfoKHR structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkWaylandSurfaceCreateInfoKHR structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkInstance, VkSurfaceKHR, VkWaylandSurfaceCreateInfoKHR

getPhysicalDeviceWaylandPresentationSupportKHR :: PhysicalDevice -> Word32 -> Ptr Wl_display -> IO VkBool32

vkGetPhysicalDeviceWaylandPresentationSupportKHR - Query physical device for presentation to Wayland

Parameters

physicalDevice is the physical device.

queueFamilyIndex is the queue family index.

display is a pointer to the wl_display associated with a Wayland compositor.

Description

This platform-specific function can be called prior to creating a surface.

Valid Usage (Implicit)

See Also

VkPhysicalDevice

pattern KHR_WAYLAND_SURFACE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_WAYLAND_SURFACE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_WAYLAND_SURFACE_CREATE_INFO_KHR :: a ~ StructureType => a