Graphics.Vulkan.Extensions.VK_KHR_win32_keyed_mutex

Documentation

withCStructWin32KeyedMutexAcquireReleaseInfoKHR :: Win32KeyedMutexAcquireReleaseInfoKHR -> (VkWin32KeyedMutexAcquireReleaseInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkWin32KeyedMutexAcquireReleaseInfoKHR and marshal a Win32KeyedMutexAcquireReleaseInfoKHR into it. The VkWin32KeyedMutexAcquireReleaseInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructWin32KeyedMutexAcquireReleaseInfoKHR :: VkWin32KeyedMutexAcquireReleaseInfoKHR -> IO Win32KeyedMutexAcquireReleaseInfoKHR

A function to read a VkWin32KeyedMutexAcquireReleaseInfoKHR and all additional structures in the pointer chain into a Win32KeyedMutexAcquireReleaseInfoKHR.

data Win32KeyedMutexAcquireReleaseInfoKHR

VkWin32KeyedMutexAcquireReleaseInfoKHR - Use the Windows keyed mutex mechanism to synchronize work

Valid Usage

Each member of pAcquireSyncs and pReleaseSyncs must be a device memory object imported by setting VkImportMemoryWin32HandleInfoKHR::handleType to VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_BIT or VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_KMT_BIT.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_WIN32_KEYED_MUTEX_ACQUIRE_RELEASE_INFO_KHR

If acquireCount is not 0, pAcquireSyncs must be a valid pointer to an array of acquireCount valid VkDeviceMemory handles

If acquireCount is not 0, pAcquireKeys must be a valid pointer to an array of acquireCount uint64_t values

If acquireCount is not 0, pAcquireTimeouts must be a valid pointer to an array of acquireCount uint32_t values

If releaseCount is not 0, pReleaseSyncs must be a valid pointer to an array of releaseCount valid VkDeviceMemory handles

If releaseCount is not 0, pReleaseKeys must be a valid pointer to an array of releaseCount uint64_t values

Both of the elements of pAcquireSyncs, and the elements of pReleaseSyncs that are valid handles must have been created, allocated, or retrieved from the same VkDevice

See Also

VkDeviceMemory, VkStructureType

Constructors

Win32KeyedMutexAcquireReleaseInfoKHR

Fields

next :: Maybe SomeVkStruct

acquireSyncs :: Vector DeviceMemory

acquireKeys :: Vector Word64

acquireTimeouts :: Vector Word32

releaseSyncs :: Vector DeviceMemory

releaseKeys :: Vector Word64

pattern KHR_WIN32_KEYED_MUTEX_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_WIN32_KEYED_MUTEX_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_WIN32_KEYED_MUTEX_ACQUIRE_RELEASE_INFO_KHR :: a ~ StructureType => a