Graphics.Vulkan.Extensions.VK_KHR_win32_surface

Documentation

createWin32SurfaceKHR :: Instance -> Win32SurfaceCreateInfoKHR -> ("allocator" ::: Maybe AllocationCallbacks) -> IO SurfaceKHR

vkCreateWin32SurfaceKHR - Create a SurfaceKHR object for an Win32 native window

Parameters

Instance is the instance to associate the surface with.

pCreateInfo is a pointer to a Win32SurfaceCreateInfoKHR structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface is a pointer to a SurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

Instance must be a valid Instance handle

pCreateInfo must be a valid pointer to a valid Win32SurfaceCreateInfoKHR structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pSurface must be a valid pointer to a SurfaceKHR handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, Instance, SurfaceKHR, Win32SurfaceCreateInfoKHR

getPhysicalDeviceWin32PresentationSupportKHR :: PhysicalDevice -> ("queueFamilyIndex" ::: Word32) -> IO Bool

vkGetPhysicalDeviceWin32PresentationSupportKHR - query queue family support for presentation on a Win32 display

Parameters

PhysicalDevice is the physical device.

queueFamilyIndex is the queue family index.

Description

This platform-specific function can be called prior to creating a surface.

Valid Usage (Implicit)

See Also

PhysicalDevice

data Win32SurfaceCreateInfoKHR

VkWin32SurfaceCreateInfoKHR - Structure specifying parameters of a newly created Win32 surface object

Valid Usage (Implicit)

See Also

StructureType, Win32SurfaceCreateFlagsKHR, createWin32SurfaceKHR

Constructors

Win32SurfaceCreateInfoKHR

Fields

flags :: Win32SurfaceCreateFlagsKHR

Flags must be 0

hinstance :: HINSTANCE

hinstance must be a valid Win32 HINSTANCE.

hwnd :: HWND

hwnd must be a valid Win32 HWND.

newtype Win32SurfaceCreateFlagsKHR

VkWin32SurfaceCreateFlagsKHR - Reserved for future use

Description

Win32SurfaceCreateFlagsKHR is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

Win32SurfaceCreateInfoKHR

Constructors

Win32SurfaceCreateFlagsKHR Flags

type KHR_WIN32_SURFACE_SPEC_VERSION = 6

pattern KHR_WIN32_SURFACE_SPEC_VERSION :: forall a. Integral a => a

type KHR_WIN32_SURFACE_EXTENSION_NAME = "VK_KHR_win32_surface"

pattern KHR_WIN32_SURFACE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype SurfaceKHR

VkSurfaceKHR - Opaque handle to a surface object

Description

The VK_KHR_surface extension declares the SurfaceKHR object, and provides a function for destroying SurfaceKHR objects. Separate platform-specific extensions each provide a function for creating a SurfaceKHR object for the respective platform. From the application’s perspective this is an opaque handle, just like the handles of other Vulkan objects.

See Also

PhysicalDeviceSurfaceInfo2KHR, SwapchainCreateInfoKHR, createAndroidSurfaceKHR, createDisplayPlaneSurfaceKHR, createHeadlessSurfaceEXT, createIOSSurfaceMVK, createImagePipeSurfaceFUCHSIA, createMacOSSurfaceMVK, createMetalSurfaceEXT, createStreamDescriptorSurfaceGGP, createViSurfaceNN, createWaylandSurfaceKHR, createWin32SurfaceKHR, createXcbSurfaceKHR, createXlibSurfaceKHR, destroySurfaceKHR, getDeviceGroupSurfacePresentModesKHR, getPhysicalDevicePresentRectanglesKHR, getPhysicalDeviceSurfaceCapabilities2EXT, getPhysicalDeviceSurfaceCapabilitiesKHR, getPhysicalDeviceSurfaceFormatsKHR, getPhysicalDeviceSurfacePresentModesKHR, getPhysicalDeviceSurfaceSupportKHR

Constructors

SurfaceKHR Word64

type HINSTANCE = Ptr ()

type HWND = Ptr ()