Graphics.Vulkan.Extensions.VK_KHR_xcb_surface

Documentation

type XcbSurfaceCreateFlagsKHR = VkXcbSurfaceCreateFlagsKHR

withCStructXcbSurfaceCreateInfoKHR :: XcbSurfaceCreateInfoKHR -> (VkXcbSurfaceCreateInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkXcbSurfaceCreateInfoKHR and marshal a XcbSurfaceCreateInfoKHR into it. The VkXcbSurfaceCreateInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructXcbSurfaceCreateInfoKHR :: VkXcbSurfaceCreateInfoKHR -> IO XcbSurfaceCreateInfoKHR

A function to read a VkXcbSurfaceCreateInfoKHR and all additional structures in the pointer chain into a XcbSurfaceCreateInfoKHR.

data XcbSurfaceCreateInfoKHR

VkXcbSurfaceCreateInfoKHR - Structure specifying parameters of a newly created Xcb surface object

Valid Usage (Implicit)

See Also

VkStructureType, VkXcbSurfaceCreateFlagsKHR, vkCreateXcbSurfaceKHR

Constructors

XcbSurfaceCreateInfoKHR

Fields

next :: Maybe SomeVkStruct

flags :: XcbSurfaceCreateFlagsKHR

connection :: Ptr Xcb_connection_t

window :: Xcb_window_t

createXcbSurfaceKHR :: Instance -> XcbSurfaceCreateInfoKHR -> Maybe AllocationCallbacks -> IO SurfaceKHR

vkCreateXcbSurfaceKHR - Create a VkSurfaceKHR object for a X11 window, using the XCB client-side library

Parameters

instance is the instance to associate the surface with.

pCreateInfo is a pointer to an instance of the VkXcbSurfaceCreateInfoKHR structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkXcbSurfaceCreateInfoKHR structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkInstance, VkSurfaceKHR, VkXcbSurfaceCreateInfoKHR

getPhysicalDeviceXcbPresentationSupportKHR :: PhysicalDevice -> Word32 -> Ptr Xcb_connection_t -> Xcb_visualid_t -> IO VkBool32

vkGetPhysicalDeviceXcbPresentationSupportKHR - Query physical device for presentation to X11 server using XCB

Parameters

physicalDevice is the physical device.

queueFamilyIndex is the queue family index.

connection is a pointer to an xcb_connection_t to the X server. visual_id is an X11 visual (xcb_visualid_t).

Description

This platform-specific function can be called prior to creating a surface.

Valid Usage (Implicit)

See Also

VkPhysicalDevice

pattern KHR_XCB_SURFACE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_XCB_SURFACE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_XCB_SURFACE_CREATE_INFO_KHR :: a ~ StructureType => a