Graphics.Vulkan.Extensions.VK_KHR_xlib_surface

Documentation

type XlibSurfaceCreateFlagsKHR = VkXlibSurfaceCreateFlagsKHR

withCStructXlibSurfaceCreateInfoKHR :: XlibSurfaceCreateInfoKHR -> (VkXlibSurfaceCreateInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkXlibSurfaceCreateInfoKHR and marshal a XlibSurfaceCreateInfoKHR into it. The VkXlibSurfaceCreateInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructXlibSurfaceCreateInfoKHR :: VkXlibSurfaceCreateInfoKHR -> IO XlibSurfaceCreateInfoKHR

A function to read a VkXlibSurfaceCreateInfoKHR and all additional structures in the pointer chain into a XlibSurfaceCreateInfoKHR.

data XlibSurfaceCreateInfoKHR

VkXlibSurfaceCreateInfoKHR - Structure specifying parameters of a newly created Xlib surface object

Valid Usage (Implicit)

See Also

VkStructureType, VkXlibSurfaceCreateFlagsKHR, vkCreateXlibSurfaceKHR

Constructors

XlibSurfaceCreateInfoKHR

Fields

next :: Maybe SomeVkStruct

flags :: XlibSurfaceCreateFlagsKHR

dpy :: Ptr Display

window :: Window

createXlibSurfaceKHR :: Instance -> XlibSurfaceCreateInfoKHR -> Maybe AllocationCallbacks -> IO SurfaceKHR

vkCreateXlibSurfaceKHR - Create a VkSurfaceKHR object for an X11 window, using the Xlib client-side library

Parameters

instance is the instance to associate the surface with.

pCreateInfo is a pointer to an instance of the VkXlibSurfaceCreateInfoKHR structure containing the parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkXlibSurfaceCreateInfoKHR structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkInstance, VkSurfaceKHR, VkXlibSurfaceCreateInfoKHR

getPhysicalDeviceXlibPresentationSupportKHR :: PhysicalDevice -> Word32 -> VisualID -> IO (VkBool32, Display)

vkGetPhysicalDeviceXlibPresentationSupportKHR - Query physical device for presentation to X11 server using Xlib

Parameters

physicalDevice is the physical device.

queueFamilyIndex is the queue family index.

dpy is a pointer to an Xlib Display connection to the server.

visualId is an X11 visual (VisualID).

Description

This platform-specific function can be called prior to creating a surface.

Valid Usage (Implicit)

See Also

VkPhysicalDevice

pattern KHR_XLIB_SURFACE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_XLIB_SURFACE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_XLIB_SURFACE_CREATE_INFO_KHR :: a ~ StructureType => a