Graphics.Vulkan.Extensions.VK_MVK_ios_surface

Documentation

type IOSSurfaceCreateFlagsMVK = VkIOSSurfaceCreateFlagsMVK

withCStructIOSSurfaceCreateInfoMVK :: IOSSurfaceCreateInfoMVK -> (VkIOSSurfaceCreateInfoMVK -> IO a) -> IO a

A function to temporarily allocate memory for a VkIOSSurfaceCreateInfoMVK and marshal a IOSSurfaceCreateInfoMVK into it. The VkIOSSurfaceCreateInfoMVK is only valid inside the provided computation and must not be returned out of it.

fromCStructIOSSurfaceCreateInfoMVK :: VkIOSSurfaceCreateInfoMVK -> IO IOSSurfaceCreateInfoMVK

A function to read a VkIOSSurfaceCreateInfoMVK and all additional structures in the pointer chain into a IOSSurfaceCreateInfoMVK.

data IOSSurfaceCreateInfoMVK

VkIOSSurfaceCreateInfoMVK - Structure specifying parameters of a newly created iOS surface object

Valid Usage (Implicit)

See Also

VkIOSSurfaceCreateFlagsMVK, VkStructureType, vkCreateIOSSurfaceMVK

Constructors

IOSSurfaceCreateInfoMVK

Fields

next :: Maybe SomeVkStruct

flags :: IOSSurfaceCreateFlagsMVK

view :: Ptr ()

createIOSSurfaceMVK :: Instance -> IOSSurfaceCreateInfoMVK -> Maybe AllocationCallbacks -> IO SurfaceKHR

vkCreateIOSSurfaceMVK - Create a VkSurfaceKHR object for an iOS UIView

Parameters

instance is the instance with which to associate the surface.

pCreateInfo is a pointer to an instance of the VkIOSSurfaceCreateInfoMVK structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkIOSSurfaceCreateInfoMVK structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_NATIVE_WINDOW_IN_USE_KHR

See Also

VkAllocationCallbacks, VkIOSSurfaceCreateInfoMVK, VkInstance, VkSurfaceKHR

pattern MVK_IOS_SURFACE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern MVK_IOS_SURFACE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_IOS_SURFACE_CREATE_INFO_MVK :: a ~ StructureType => a