Graphics.Vulkan.Extensions.VK_NN_vi_surface

Documentation

type ViSurfaceCreateFlagsNN = VkViSurfaceCreateFlagsNN

withCStructViSurfaceCreateInfoNN :: ViSurfaceCreateInfoNN -> (VkViSurfaceCreateInfoNN -> IO a) -> IO a

A function to temporarily allocate memory for a VkViSurfaceCreateInfoNN and marshal a ViSurfaceCreateInfoNN into it. The VkViSurfaceCreateInfoNN is only valid inside the provided computation and must not be returned out of it.

fromCStructViSurfaceCreateInfoNN :: VkViSurfaceCreateInfoNN -> IO ViSurfaceCreateInfoNN

A function to read a VkViSurfaceCreateInfoNN and all additional structures in the pointer chain into a ViSurfaceCreateInfoNN.

data ViSurfaceCreateInfoNN

VkViSurfaceCreateInfoNN - Structure specifying parameters of a newly created VI surface object

Valid Usage (Implicit)

See Also

VkStructureType, VkViSurfaceCreateFlagsNN, vkCreateViSurfaceNN

Constructors

ViSurfaceCreateInfoNN

Fields

next :: Maybe SomeVkStruct

flags :: ViSurfaceCreateFlagsNN

window :: Ptr ()

createViSurfaceNN :: Instance -> ViSurfaceCreateInfoNN -> Maybe AllocationCallbacks -> IO SurfaceKHR

vkCreateViSurfaceNN - Create a VkSurfaceKHR object for a VI layer

Parameters

instance is the instance with which to associate the surface.

pCreateInfo is a pointer to an instance of the VkViSurfaceCreateInfoNN structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface object is returned.

Description

During the lifetime of a surface created using a particular nn::vi::NativeWindowHandle, applications must not attempt to create another surface for the same nn::vi::Layer or attempt to connect to the same nn::vi::Layer through other platform mechanisms.

If the native window is created with a specified size, currentExtent will reflect that size. In this case, applications should use the same size for the swapchain’s imageExtent. Otherwise, the currentExtent will have the special value (0xFFFFFFFF, 0xFFFFFFFF), indicating that applications are expected to choose an appropriate size for the swapchain’s imageExtent (e.g., by matching the result of a call to nn::vi::GetDisplayResolution).

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkViSurfaceCreateInfoNN structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_NATIVE_WINDOW_IN_USE_KHR

See Also

VkAllocationCallbacks, VkInstance, VkSurfaceKHR, VkViSurfaceCreateInfoNN

pattern NN_VI_SURFACE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NN_VI_SURFACE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_VI_SURFACE_CREATE_INFO_NN :: a ~ StructureType => a