Graphics.Vulkan.Extensions.VK_NVX_device_generated_commands

Documentation

cmdProcessCommandsNVX :: CommandBuffer -> CmdProcessCommandsInfoNVX -> IO ()

vkCmdProcessCommandsNVX - Performs the generation of commands on the device

Parameters

CommandBuffer is the primary command buffer in which the generation process takes space.

pProcessCommandsInfo is a pointer to a CmdProcessCommandsInfoNVX structure containing parameters affecting the processing of commands.

Valid Usage (Implicit)

CommandBuffer must be a valid CommandBuffer handle

pProcessCommandsInfo must be a valid pointer to a valid CmdProcessCommandsInfoNVX structure

CommandBuffer must be in the recording state

The CommandPool that CommandBuffer was allocated from must support graphics, or compute operations

This command must only be called inside of a render pass instance

Host Synchronization

Host access to CommandBuffer must be externally synchronized

Host access to the CommandPool that CommandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Inside 	 Graphics Compute

See Also

CmdProcessCommandsInfoNVX, CommandBuffer

cmdReserveSpaceForCommandsNVX :: CommandBuffer -> ("reserveSpaceInfo" ::: CmdReserveSpaceForCommandsInfoNVX) -> IO ()

vkCmdReserveSpaceForCommandsNVX - Perform a reservation of command buffer space

Parameters

CommandBuffer is the secondary command buffer in which the space for device-generated commands is reserved.

pProcessCommandsInfo is a pointer to a CmdReserveSpaceForCommandsInfoNVX structure containing parameters affecting the reservation of command buffer space.

Valid Usage

The provided CommandBuffer must not have had a prior space reservation since its creation or the last reset.

The state of the CommandBuffer must be legal to execute all commands within the sequence provided by the indirectCommandsLayout member of pProcessCommandsInfo.

Valid Usage (Implicit)

CommandBuffer must be a valid CommandBuffer handle

pReserveSpaceInfo must be a valid pointer to a valid CmdReserveSpaceForCommandsInfoNVX structure

CommandBuffer must be in the recording state

The CommandPool that CommandBuffer was allocated from must support graphics, or compute operations

This command must only be called inside of a render pass instance

CommandBuffer must be a secondary CommandBuffer

Host Synchronization

Host access to CommandBuffer must be externally synchronized

Host access to the CommandPool that CommandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Secondary 	 Inside 	 Graphics Compute

See Also

CmdReserveSpaceForCommandsInfoNVX, CommandBuffer

createIndirectCommandsLayoutNVX :: Device -> IndirectCommandsLayoutCreateInfoNVX -> ("allocator" ::: Maybe AllocationCallbacks) -> IO IndirectCommandsLayoutNVX

vkCreateIndirectCommandsLayoutNVX - Create an indirect command layout object

Parameters

Device is the logical device that creates the indirect command layout.

pCreateInfo is a pointer to a IndirectCommandsLayoutCreateInfoNVX structure containing parameters affecting creation of the indirect command layout.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pIndirectCommandsLayout is a pointer to a IndirectCommandsLayoutNVX handle in which the resulting indirect command layout is returned.

Valid Usage (Implicit)

Device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid IndirectCommandsLayoutCreateInfoNVX structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pIndirectCommandsLayout must be a valid pointer to a IndirectCommandsLayoutNVX handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, Device, IndirectCommandsLayoutCreateInfoNVX, IndirectCommandsLayoutNVX

withIndirectCommandsLayoutNVX :: Device -> IndirectCommandsLayoutCreateInfoNVX -> Maybe AllocationCallbacks -> (IndirectCommandsLayoutNVX -> IO r) -> IO r

A safe wrapper for createIndirectCommandsLayoutNVX and destroyIndirectCommandsLayoutNVX using bracket

The allocated value must not be returned from the provided computation

destroyIndirectCommandsLayoutNVX :: Device -> IndirectCommandsLayoutNVX -> ("allocator" ::: Maybe AllocationCallbacks) -> IO ()

vkDestroyIndirectCommandsLayoutNVX - Destroy an object table

Parameters

Device is the logical device that destroys the layout.

indirectCommandsLayout is the table to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to indirectCommandsLayout must have completed execution

If AllocationCallbacks were provided when objectTable was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when objectTable was created, pAllocator must be NULL

Valid Usage (Implicit)

Device must be a valid Device handle

indirectCommandsLayout must be a valid IndirectCommandsLayoutNVX handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

indirectCommandsLayout must have been created, allocated, or retrieved from Device

See Also

AllocationCallbacks, Device, IndirectCommandsLayoutNVX

createObjectTableNVX :: Device -> ObjectTableCreateInfoNVX -> ("allocator" ::: Maybe AllocationCallbacks) -> IO ObjectTableNVX

vkCreateObjectTableNVX - Create an object table

Parameters

Device is the logical device that creates the object table.

pCreateInfo is a pointer to a ObjectTableCreateInfoNVX structure containing parameters affecting creation of the table.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pObjectTable is a pointer to a ObjectTableNVX handle in which the resulting object table is returned.

Valid Usage (Implicit)

Device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid ObjectTableCreateInfoNVX structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pObjectTable must be a valid pointer to a ObjectTableNVX handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, Device, ObjectTableCreateInfoNVX, ObjectTableNVX

withObjectTableNVX :: Device -> ObjectTableCreateInfoNVX -> Maybe AllocationCallbacks -> (ObjectTableNVX -> IO r) -> IO r

A safe wrapper for createObjectTableNVX and destroyObjectTableNVX using bracket

The allocated value must not be returned from the provided computation

destroyObjectTableNVX :: Device -> ObjectTableNVX -> ("allocator" ::: Maybe AllocationCallbacks) -> IO ()

vkDestroyObjectTableNVX - Destroy an object table

Parameters

Device is the logical device that destroys the table.

objectTable is the table to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to objectTable must have completed execution.

If AllocationCallbacks were provided when objectTable was created, a compatible set of callbacks must be provided here.

If no AllocationCallbacks were provided when objectTable was created, pAllocator must be NULL.

Valid Usage (Implicit)

Device must be a valid Device handle

objectTable must be a valid ObjectTableNVX handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

objectTable must have been created, allocated, or retrieved from Device

Host Synchronization

Host access to objectTable must be externally synchronized

See Also

AllocationCallbacks, Device, ObjectTableNVX

registerObjectsNVX :: Device -> ObjectTableNVX -> ("objectTableEntries" ::: Vector ObjectTableEntryNVX) -> ("objectIndices" ::: Vector Word32) -> IO ()

vkRegisterObjectsNVX - Register resource bindings in an object table

Parameters

Device is the logical device that creates the object table.

objectTable is the table for which the resources are registered.

objectCount is the number of resources to register.

ppObjectTableEntries provides an array for detailed binding informations. Each array element is a pointer to a structure of type ObjectTablePipelineEntryNVX, ObjectTableDescriptorSetEntryNVX, ObjectTableVertexBufferEntryNVX, ObjectTableIndexBufferEntryNVX or ObjectTablePushConstantEntryNVX (see below for details).

pObjectIndices are the indices at which each resource is registered.

Valid Usage

The contents of pObjectTableEntry must yield plausible bindings supported by the device.

At any pObjectIndices there must not be a registered resource already.

Any value inside pObjectIndices must be below the appropriate ObjectTableCreateInfoNVX::pObjectEntryCounts limits provided at objectTable creation time.

Valid Usage (Implicit)

Device must be a valid Device handle

objectTable must be a valid ObjectTableNVX handle

ppObjectTableEntries must be a valid pointer to an array of objectCount valid ObjectTableEntryNVX structures

pObjectIndices must be a valid pointer to an array of objectCount uint32_t values

objectCount must be greater than 0

objectTable must have been created, allocated, or retrieved from Device

Host Synchronization

Host access to objectTable must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

Device, ObjectTableEntryNVX, ObjectTableNVX

withRegisteredObjectsNVX :: Device -> ObjectTableNVX -> Vector ObjectTableEntryNVX -> Vector Word32 -> Vector ObjectEntryTypeNVX -> IO r -> IO r

A safe wrapper for registerObjectsNVX and unregisterObjectsNVX using bracket_

The allocated value must not be returned from the provided computation

unregisterObjectsNVX :: Device -> ObjectTableNVX -> ("objectEntryTypes" ::: Vector ObjectEntryTypeNVX) -> ("objectIndices" ::: Vector Word32) -> IO ()

vkUnregisterObjectsNVX - Unregister resource bindings in an object table

Parameters

Device is the logical device that creates the object table.

objectTable is the table from which the resources are unregistered.

objectCount is the number of resources being removed from the object table.

pObjectEntryType provides an array of ObjectEntryTypeNVX for the resources being removed.

pObjectIndices provides the array of object indices to be removed.

Valid Usage

At any pObjectIndices there must be a registered resource already.

The pObjectEntryTypes of the resource at pObjectIndices must match.

All operations on the device using the registered resource must have been completed.

Valid Usage (Implicit)

Device must be a valid Device handle

objectTable must be a valid ObjectTableNVX handle

pObjectEntryTypes must be a valid pointer to an array of objectCount valid ObjectEntryTypeNVX values

pObjectIndices must be a valid pointer to an array of objectCount uint32_t values

objectCount must be greater than 0

objectTable must have been created, allocated, or retrieved from Device

Host Synchronization

Host access to objectTable must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

Device, ObjectEntryTypeNVX, ObjectTableNVX

getPhysicalDeviceGeneratedCommandsPropertiesNVX :: PhysicalDevice -> IO (DeviceGeneratedCommandsFeaturesNVX, DeviceGeneratedCommandsLimitsNVX)

vkGetPhysicalDeviceGeneratedCommandsPropertiesNVX - Returns device-generated commands related properties of a physical device

Parameters

PhysicalDevice is the handle to the physical device whose properties will be queried.

pFeatures is a pointer to a DeviceGeneratedCommandsFeaturesNVX structure in which features are returned.

pLimits is a pointer to a DeviceGeneratedCommandsLimitsNVX structure in which limitations are returned.

Valid Usage (Implicit)

See Also

DeviceGeneratedCommandsFeaturesNVX, DeviceGeneratedCommandsLimitsNVX, PhysicalDevice

data DeviceGeneratedCommandsFeaturesNVX

VkDeviceGeneratedCommandsFeaturesNVX - Structure specifying physical device support

Valid Usage (Implicit)

See Also

Bool32, StructureType, getPhysicalDeviceGeneratedCommandsPropertiesNVX

Constructors

DeviceGeneratedCommandsFeaturesNVX

Fields

computeBindingPointSupport :: Bool

computeBindingPointSupport specifies whether the ObjectTableNVX supports entries with OBJECT_ENTRY_USAGE_GRAPHICS_BIT_NVX bit set and IndirectCommandsLayoutNVX supports PIPELINE_BIND_POINT_COMPUTE.

data DeviceGeneratedCommandsLimitsNVX

VkDeviceGeneratedCommandsLimitsNVX - Structure specifying physical device limits

Valid Usage (Implicit)

See Also

StructureType, getPhysicalDeviceGeneratedCommandsPropertiesNVX

Constructors

DeviceGeneratedCommandsLimitsNVX

Fields

maxIndirectCommandsLayoutTokenCount :: Word32

maxIndirectCommandsLayoutTokenCount the maximum number of tokens in IndirectCommandsLayoutNVX.

maxObjectEntryCounts :: Word32

maxObjectEntryCounts the maximum number of entries per resource type in ObjectTableNVX.

minSequenceCountBufferOffsetAlignment :: Word32

minSequenceCountBufferOffsetAlignment the minimum alignment for memory addresses optionally used in cmdProcessCommandsNVX.

minSequenceIndexBufferOffsetAlignment :: Word32

minSequenceIndexBufferOffsetAlignment the minimum alignment for memory addresses optionally used in cmdProcessCommandsNVX.

minCommandsTokenBufferOffsetAlignment :: Word32

minCommandsTokenBufferOffsetAlignment the minimum alignment for memory addresses optionally used in cmdProcessCommandsNVX.

data IndirectCommandsTokenNVX

VkIndirectCommandsTokenNVX - Structure specifying parameters for the reservation of command buffer space

Valid Usage

The Buffer’s usage flag must have the BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set.

The offset must be aligned to DeviceGeneratedCommandsLimitsNVX::minCommandsTokenBufferOffsetAlignment.

Valid Usage (Implicit)

tokenType must be a valid IndirectCommandsTokenTypeNVX value

Buffer must be a valid Buffer handle

See Also

Buffer, CmdProcessCommandsInfoNVX, DeviceSize, IndirectCommandsTokenTypeNVX

Constructors

IndirectCommandsTokenNVX

Fields

tokenType :: IndirectCommandsTokenTypeNVX

tokenType specifies the token command type.

buffer :: Buffer

Buffer specifies the Buffer storing the functional arguments for each squence. These argumetns can be written by the device.

offset :: DeviceSize

offset specified an offset into Buffer where the arguments start.

data IndirectCommandsLayoutTokenNVX

VkIndirectCommandsLayoutTokenNVX - Struct specifying the details of an indirect command layout token

Valid Usage (Implicit)

See Also

IndirectCommandsLayoutCreateInfoNVX, IndirectCommandsTokenTypeNVX

Constructors

IndirectCommandsLayoutTokenNVX

Fields

tokenType :: IndirectCommandsTokenTypeNVX

tokenType must be a valid IndirectCommandsTokenTypeNVX value

bindingUnit :: Word32

bindingUnit must stay within device supported limits for the appropriate commands.

dynamicCount :: Word32

dynamicCount must stay within device supported limits for the appropriate commands.

divisor :: Word32

divisor must be greater than 0 and a power of two.

data IndirectCommandsLayoutCreateInfoNVX

VkIndirectCommandsLayoutCreateInfoNVX - Structure specifying the parameters of a newly created indirect commands layout object

Description

The following code illustrates some of the key flags:

void cmdProcessAllSequences(cmd, objectTable, indirectCommandsLayout, pIndirectCommandsTokens, sequencesCount, indexbuffer, indexbufferoffset)
{
  for (s = 0; s < sequencesCount; s++)
  {
    sequence = s;

    if (indirectCommandsLayout.flags & VK_INDIRECT_COMMANDS_LAYOUT_USAGE_UNORDERED_SEQUENCES_BIT_NVX) {
      sequence = incoherent_implementation_dependent_permutation[ sequence ];
    }
    if (indirectCommandsLayout.flags & VK_INDIRECT_COMMANDS_LAYOUT_USAGE_INDEXED_SEQUENCES_BIT_NVX) {
      sequence = indexbuffer.load_uint32( sequence * sizeof(uint32_t) + indexbufferoffset);
    }

    cmdProcessSequence( cmd, objectTable, indirectCommandsLayout, pIndirectCommandsTokens, sequence );
  }
}

Valid Usage

tokenCount must be greater than 0 and below DeviceGeneratedCommandsLimitsNVX::maxIndirectCommandsLayoutTokenCount

If the DeviceGeneratedCommandsFeaturesNVX::computeBindingPointSupport feature is not enabled, then PipelineBindPoint must not be PIPELINE_BIND_POINT_COMPUTE

If pTokens contains an entry of INDIRECT_COMMANDS_TOKEN_TYPE_PIPELINE_NVX it must be the first element of the array and there must be only a single element of such token type.

All state binding tokens in pTokens must occur prior work provoking tokens (INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_NVX, INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_INDEXED_NVX, INDIRECT_COMMANDS_TOKEN_TYPE_DISPATCH_NVX).

The content of pTokens must include one single work provoking token that is compatible with the PipelineBindPoint.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_INDIRECT_COMMANDS_LAYOUT_CREATE_INFO_NVX

pNext must be NULL

PipelineBindPoint must be a valid PipelineBindPoint value

Flags must be a valid combination of IndirectCommandsLayoutUsageFlagBitsNVX values

Flags must not be 0

pTokens must be a valid pointer to an array of tokenCount valid IndirectCommandsLayoutTokenNVX structures

tokenCount must be greater than 0

See Also

IndirectCommandsLayoutTokenNVX, IndirectCommandsLayoutUsageFlagsNVX, PipelineBindPoint, StructureType, createIndirectCommandsLayoutNVX

Constructors

IndirectCommandsLayoutCreateInfoNVX

Fields

pipelineBindPoint :: PipelineBindPoint

PipelineBindPoint is the PipelineBindPoint that this layout targets.

flags :: IndirectCommandsLayoutUsageFlagsNVX

Flags is a bitmask of IndirectCommandsLayoutUsageFlagBitsNVX specifying usage hints of this layout.

tokens :: Vector IndirectCommandsLayoutTokenNVX

pTokens is an array describing each command token in detail. See IndirectCommandsTokenTypeNVX and IndirectCommandsLayoutTokenNVX below for details.

data CmdProcessCommandsInfoNVX

VkCmdProcessCommandsInfoNVX - Structure specifying parameters for the generation of commands

Valid Usage

The provided objectTable must include all objects referenced by the generation process

indirectCommandsTokenCount must match the indirectCommandsLayout’s tokenCount

The tokenType member of each entry in the pIndirectCommandsTokens array must match the values used at creation time of indirectCommandsLayout

If targetCommandBuffer is provided, it must have reserved command space

If targetCommandBuffer is provided, the objectTable must match the reservation’s objectTable and must have had all referenced objects registered at reservation time

If targetCommandBuffer is provided, the indirectCommandsLayout must match the reservation’s indirectCommandsLayout

If targetCommandBuffer is provided, the maxSequencesCount must not exceed the reservation’s maxSequencesCount

If sequencesCountBuffer is used, its usage flag must have the BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set

If sequencesCountBuffer is used, sequencesCountOffset must be aligned to DeviceGeneratedCommandsLimitsNVX::minSequenceCountBufferOffsetAlignment

If sequencesIndexBuffer is used, its usage flag must have the BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set

If sequencesIndexBuffer is used, sequencesIndexOffset must be aligned to DeviceGeneratedCommandsLimitsNVX::minSequenceIndexBufferOffsetAlignment

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_CMD_PROCESS_COMMANDS_INFO_NVX

pNext must be NULL

objectTable must be a valid ObjectTableNVX handle

indirectCommandsLayout must be a valid IndirectCommandsLayoutNVX handle

pIndirectCommandsTokens must be a valid pointer to an array of indirectCommandsTokenCount valid IndirectCommandsTokenNVX structures

If targetCommandBuffer is not NULL, targetCommandBuffer must be a valid CommandBuffer handle

If sequencesCountBuffer is not NULL_HANDLE, sequencesCountBuffer must be a valid Buffer handle

If sequencesIndexBuffer is not NULL_HANDLE, sequencesIndexBuffer must be a valid Buffer handle

indirectCommandsTokenCount must be greater than 0

Each of indirectCommandsLayout, objectTable, sequencesCountBuffer, sequencesIndexBuffer, and targetCommandBuffer that are valid handles of non-ignored parameters must have been created, allocated, or retrieved from the same Device

Host Synchronization

Host access to objectTable must be externally synchronized

Host access to targetCommandBuffer must be externally synchronized

See Also

Buffer, CommandBuffer, DeviceSize, IndirectCommandsLayoutNVX, IndirectCommandsTokenNVX, ObjectTableNVX, StructureType, cmdProcessCommandsNVX

Constructors

CmdProcessCommandsInfoNVX

Fields

objectTable :: ObjectTableNVX

objectTable is the ObjectTableNVX to be used for the generation process. Only registered objects at the time cmdReserveSpaceForCommandsNVX is called, will be taken into account for the reservation.

indirectCommandsLayout :: IndirectCommandsLayoutNVX

indirectCommandsLayout is the IndirectCommandsLayoutNVX that provides the command sequence to generate.

indirectCommandsTokens :: Vector IndirectCommandsTokenNVX

pIndirectCommandsTokens provides an array of IndirectCommandsTokenNVX that reference the input data for each token command.

maxSequencesCount :: Word32

maxSequencesCount is the maximum number of sequences for which command buffer space will be reserved. If sequencesCountBuffer is NULL_HANDLE, this is also the actual number of sequences generated.

targetCommandBuffer :: Ptr CommandBuffer_T

targetCommandBuffer can be the secondary CommandBuffer in which the commands should be recorded. If targetCommandBuffer is NULL an implicit reservation as well as execution takes place on the processing CommandBuffer.

sequencesCountBuffer :: Buffer

sequencesCountBuffer can be Buffer from which the actual amount of sequences is sourced from as uint32_t value.

sequencesCountOffset :: DeviceSize

sequencesCountOffset is the byte offset into sequencesCountBuffer where the count value is stored.

sequencesIndexBuffer :: Buffer

sequencesIndexBuffer must be set if indirectCommandsLayout’s INDIRECT_COMMANDS_LAYOUT_USAGE_INDEXED_SEQUENCES_BIT_NVX is set and provides the used sequence indices as uint32_t array. Otherwise it must be NULL_HANDLE.

sequencesIndexOffset :: DeviceSize

sequencesIndexOffset is the byte offset into sequencesIndexBuffer where the index values start.

data CmdReserveSpaceForCommandsInfoNVX

VkCmdReserveSpaceForCommandsInfoNVX - Structure specifying parameters for the reservation of command buffer space

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_CMD_RESERVE_SPACE_FOR_COMMANDS_INFO_NVX

pNext must be NULL

objectTable must be a valid ObjectTableNVX handle

indirectCommandsLayout must be a valid IndirectCommandsLayoutNVX handle

Both of indirectCommandsLayout, and objectTable must have been created, allocated, or retrieved from the same Device

Host Synchronization

Host access to objectTable must be externally synchronized

See Also

IndirectCommandsLayoutNVX, ObjectTableNVX, StructureType, cmdReserveSpaceForCommandsNVX

Constructors

CmdReserveSpaceForCommandsInfoNVX

Fields

objectTable :: ObjectTableNVX

objectTable is the ObjectTableNVX to be used for the generation process. Only registered objects at the time cmdReserveSpaceForCommandsNVX is called, will be taken into account for the reservation.

indirectCommandsLayout :: IndirectCommandsLayoutNVX

indirectCommandsLayout is the IndirectCommandsLayoutNVX that must also be used at generation time.

maxSequencesCount :: Word32

maxSequencesCount is the maximum number of sequences for which command buffer space will be reserved.

data ObjectTableCreateInfoNVX

VkObjectTableCreateInfoNVX - Structure specifying the parameters of a newly created object table

Valid Usage

If the DeviceGeneratedCommandsFeaturesNVX::computeBindingPointSupport feature is not enabled, pObjectEntryUsageFlags must not contain OBJECT_ENTRY_USAGE_COMPUTE_BIT_NVX

Any value within pObjectEntryCounts must not exceed DeviceGeneratedCommandsLimitsNVX::maxObjectEntryCounts

maxUniformBuffersPerDescriptor must be within the limits supported by the device.

maxStorageBuffersPerDescriptor must be within the limits supported by the device.

maxStorageImagesPerDescriptor must be within the limits supported by the device.

maxSampledImagesPerDescriptor must be within the limits supported by the device.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_OBJECT_TABLE_CREATE_INFO_NVX

pNext must be NULL

pObjectEntryTypes must be a valid pointer to an array of objectCount valid ObjectEntryTypeNVX values

pObjectEntryCounts must be a valid pointer to an array of objectCount uint32_t values

pObjectEntryUsageFlags must be a valid pointer to an array of objectCount valid combinations of ObjectEntryUsageFlagBitsNVX values

Each element of pObjectEntryUsageFlags must not be 0

objectCount must be greater than 0

See Also

ObjectEntryTypeNVX, ObjectEntryUsageFlagsNVX, StructureType, createObjectTableNVX

Constructors

ObjectTableCreateInfoNVX

Fields

objectEntryTypes :: Vector ObjectEntryTypeNVX

pObjectEntryTypes is a pointer to an array of ObjectEntryTypeNVX values providing the entry type of a given configuration.

objectEntryCounts :: Vector Word32

pObjectEntryCounts is a pointer to an array of counts of how many objects can be registered in the table.

objectEntryUsageFlags :: Vector ObjectEntryUsageFlagsNVX

pObjectEntryUsageFlags is a pointer to an array of bitmasks of ObjectEntryUsageFlagBitsNVX specifying the binding usage of the entry.

maxUniformBuffersPerDescriptor :: Word32

maxUniformBuffersPerDescriptor is the maximum number of DESCRIPTOR_TYPE_UNIFORM_BUFFER or DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC used by any single registered DescriptorSet in this table.

maxStorageBuffersPerDescriptor :: Word32

maxStorageBuffersPerDescriptor is the maximum number of DESCRIPTOR_TYPE_STORAGE_BUFFER or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC used by any single registered DescriptorSet in this table.

maxStorageImagesPerDescriptor :: Word32

maxStorageImagesPerDescriptor is the maximum number of DESCRIPTOR_TYPE_STORAGE_IMAGE or DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER used by any single registered DescriptorSet in this table.

maxSampledImagesPerDescriptor :: Word32

maxSampledImagesPerDescriptor is the maximum number of DESCRIPTOR_TYPE_SAMPLER, DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER or DESCRIPTOR_TYPE_INPUT_ATTACHMENT used by any single registered DescriptorSet in this table.

maxPipelineLayouts :: Word32

maxPipelineLayouts is the maximum number of unique PipelineLayout used by any registered DescriptorSet or Pipeline in this table.

data ObjectTableEntryNVX

VkObjectTableEntryNVX - Common parameters of an object table resource entry

Valid Usage

If the DeviceGeneratedCommandsFeaturesNVX::computeBindingPointSupport feature is not enabled, Flags must not contain OBJECT_ENTRY_USAGE_COMPUTE_BIT_NVX

Valid Usage (Implicit)

type must be a valid ObjectEntryTypeNVX value

Flags must be a valid combination of ObjectEntryUsageFlagBitsNVX values

Flags must not be 0

See Also

ObjectEntryTypeNVX, ObjectEntryUsageFlagsNVX, registerObjectsNVX

Constructors

ObjectTableEntryNVX

Fields

type' :: ObjectEntryTypeNVX

type defines the entry type

flags :: ObjectEntryUsageFlagsNVX

Flags defines which PipelineBindPoint the resource can be used with. Some entry types allow only a single flag to be set.

data ObjectTablePipelineEntryNVX

VkObjectTablePipelineEntryNVX - Parameters of an object table pipeline entry

Valid Usage (Implicit)

See Also

ObjectEntryTypeNVX, ObjectEntryUsageFlagsNVX, Pipeline

Constructors

ObjectTablePipelineEntryNVX

Fields

type' :: ObjectEntryTypeNVX

type must be a valid ObjectEntryTypeNVX value

flags :: ObjectEntryUsageFlagsNVX

Flags must not be 0

pipeline :: Pipeline

Pipeline must be a valid Pipeline handle

data ObjectTableDescriptorSetEntryNVX

VkObjectTableDescriptorSetEntryNVX - Parameters of an object table descriptor set entry

Valid Usage

type must be OBJECT_ENTRY_TYPE_DESCRIPTOR_SET_NVX

Valid Usage (Implicit)

type must be a valid ObjectEntryTypeNVX value

Flags must be a valid combination of ObjectEntryUsageFlagBitsNVX values

Flags must not be 0

PipelineLayout must be a valid PipelineLayout handle

DescriptorSet must be a valid DescriptorSet handle

Both of DescriptorSet, and PipelineLayout must have been created, allocated, or retrieved from the same Device

See Also

DescriptorSet, ObjectEntryTypeNVX, ObjectEntryUsageFlagsNVX, PipelineLayout

Constructors

ObjectTableDescriptorSetEntryNVX

Fields

type' :: ObjectEntryTypeNVX

flags :: ObjectEntryUsageFlagsNVX

pipelineLayout :: PipelineLayout

PipelineLayout specifies the PipelineLayout that the DescriptorSet is used with.

descriptorSet :: DescriptorSet

DescriptorSet specifies the DescriptorSet that can be bound with this entry.

data ObjectTableVertexBufferEntryNVX

VkObjectTableVertexBufferEntryNVX - Parameters of an object table vertex buffer entry

Valid Usage (Implicit)

See Also

Buffer, ObjectEntryTypeNVX, ObjectEntryUsageFlagsNVX

Constructors

ObjectTableVertexBufferEntryNVX

Fields

type' :: ObjectEntryTypeNVX

type must be a valid ObjectEntryTypeNVX value

flags :: ObjectEntryUsageFlagsNVX

Flags must not be 0

buffer :: Buffer

Buffer must be a valid Buffer handle

data ObjectTableIndexBufferEntryNVX

VkObjectTableIndexBufferEntryNVX - Parameters of an object table index buffer entry

Valid Usage (Implicit)

See Also

Buffer, IndexType, ObjectEntryTypeNVX, ObjectEntryUsageFlagsNVX

Constructors

ObjectTableIndexBufferEntryNVX

Fields

type' :: ObjectEntryTypeNVX

type must be a valid ObjectEntryTypeNVX value

flags :: ObjectEntryUsageFlagsNVX

Flags must not be 0

buffer :: Buffer

Buffer must be a valid Buffer handle

indexType :: IndexType

IndexType must be a valid IndexType value

data ObjectTablePushConstantEntryNVX

VkObjectTablePushConstantEntryNVX - Parameters of an object table push constant entry

Valid Usage (Implicit)

See Also

ObjectEntryTypeNVX, ObjectEntryUsageFlagsNVX, PipelineLayout, ShaderStageFlags

Constructors

ObjectTablePushConstantEntryNVX

Fields

type' :: ObjectEntryTypeNVX

type must be a valid ObjectEntryTypeNVX value

flags :: ObjectEntryUsageFlagsNVX

Flags must not be 0

pipelineLayout :: PipelineLayout

PipelineLayout must be a valid PipelineLayout handle

stageFlags :: ShaderStageFlags

stageFlags must not be 0

newtype IndirectCommandsLayoutUsageFlagBitsNVX

VkIndirectCommandsLayoutUsageFlagBitsNVX - Bitmask specifying allowed usage of an indirect commands layout

See Also

IndirectCommandsLayoutUsageFlagsNVX

Constructors

IndirectCommandsLayoutUsageFlagBitsNVX Flags

Bundled Patterns

pattern INDIRECT_COMMANDS_LAYOUT_USAGE_UNORDERED_SEQUENCES_BIT_NVX :: IndirectCommandsLayoutUsageFlagBitsNVX
INDIRECT_COMMANDS_LAYOUT_USAGE_UNORDERED_SEQUENCES_BIT_NVX specifies that the processing of sequences can happen at an implementation-dependent order, which is not guaranteed to be coherent across multiple invocations.

pattern INDIRECT_COMMANDS_LAYOUT_USAGE_SPARSE_SEQUENCES_BIT_NVX :: IndirectCommandsLayoutUsageFlagBitsNVX
INDIRECT_COMMANDS_LAYOUT_USAGE_SPARSE_SEQUENCES_BIT_NVX specifies that there is likely a high difference between allocated number of sequences and actually used.

pattern INDIRECT_COMMANDS_LAYOUT_USAGE_EMPTY_EXECUTIONS_BIT_NVX :: IndirectCommandsLayoutUsageFlagBitsNVX
INDIRECT_COMMANDS_LAYOUT_USAGE_EMPTY_EXECUTIONS_BIT_NVX specifies that there are likely many draw or dispatch calls that are zero-sized (zero grid dimension, no primitives to render).

pattern INDIRECT_COMMANDS_LAYOUT_USAGE_INDEXED_SEQUENCES_BIT_NVX :: IndirectCommandsLayoutUsageFlagBitsNVX
INDIRECT_COMMANDS_LAYOUT_USAGE_INDEXED_SEQUENCES_BIT_NVX specifies that the input data for the sequences is not implicitly indexed from 0..sequencesUsed but a user provided Buffer encoding the index is provided.

type IndirectCommandsLayoutUsageFlagsNVX = IndirectCommandsLayoutUsageFlagBitsNVX

newtype ObjectEntryUsageFlagBitsNVX

VkObjectEntryUsageFlagBitsNVX - Bitmask specifying allowed usage of an object entry

See Also

ObjectEntryUsageFlagsNVX

Constructors

ObjectEntryUsageFlagBitsNVX Flags

Bundled Patterns

pattern OBJECT_ENTRY_USAGE_GRAPHICS_BIT_NVX :: ObjectEntryUsageFlagBitsNVX
OBJECT_ENTRY_USAGE_GRAPHICS_BIT_NVX specifies that the resource is bound to PIPELINE_BIND_POINT_GRAPHICS

pattern OBJECT_ENTRY_USAGE_COMPUTE_BIT_NVX :: ObjectEntryUsageFlagBitsNVX
OBJECT_ENTRY_USAGE_COMPUTE_BIT_NVX specifies that the resource is bound to PIPELINE_BIND_POINT_COMPUTE

type ObjectEntryUsageFlagsNVX = ObjectEntryUsageFlagBitsNVX

newtype IndirectCommandsTokenTypeNVX

VkIndirectCommandsTokenTypeNVX - Enum specifying

Description

'

Token type 	 Equivalent command

INDIRECT_COMMANDS_TOKEN_TYPE_PIPELINE_NVX 	 cmdBindPipeline

INDIRECT_COMMANDS_TOKEN_TYPE_DESCRIPTOR_SET_NVX 	 cmdBindDescriptorSets

INDIRECT_COMMANDS_TOKEN_TYPE_INDEX_BUFFER_NVX 	 cmdBindIndexBuffer

INDIRECT_COMMANDS_TOKEN_TYPE_VERTEX_BUFFER_NVX 	 cmdBindVertexBuffers

INDIRECT_COMMANDS_TOKEN_TYPE_PUSH_CONSTANT_NVX 	 cmdPushConstants

INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_INDEXED_NVX 	 cmdDrawIndexedIndirect

INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_NVX 	 cmdDrawIndirect

INDIRECT_COMMANDS_TOKEN_TYPE_DISPATCH_NVX 	 cmdDispatchIndirect

Supported indirect command tokens

See Also

IndirectCommandsLayoutTokenNVX, IndirectCommandsTokenNVX

Constructors

IndirectCommandsTokenTypeNVX Int32

Bundled Patterns

pattern INDIRECT_COMMANDS_TOKEN_TYPE_PIPELINE_NVX :: IndirectCommandsTokenTypeNVX

pattern INDIRECT_COMMANDS_TOKEN_TYPE_DESCRIPTOR_SET_NVX :: IndirectCommandsTokenTypeNVX

pattern INDIRECT_COMMANDS_TOKEN_TYPE_INDEX_BUFFER_NVX :: IndirectCommandsTokenTypeNVX

pattern INDIRECT_COMMANDS_TOKEN_TYPE_VERTEX_BUFFER_NVX :: IndirectCommandsTokenTypeNVX

pattern INDIRECT_COMMANDS_TOKEN_TYPE_PUSH_CONSTANT_NVX :: IndirectCommandsTokenTypeNVX

pattern INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_INDEXED_NVX :: IndirectCommandsTokenTypeNVX

pattern INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_NVX :: IndirectCommandsTokenTypeNVX

pattern INDIRECT_COMMANDS_TOKEN_TYPE_DISPATCH_NVX :: IndirectCommandsTokenTypeNVX

newtype ObjectEntryTypeNVX

VkObjectEntryTypeNVX - Enum specifying object table entry type

See Also

ObjectTableCreateInfoNVX, ObjectTableDescriptorSetEntryNVX, ObjectTableEntryNVX, ObjectTableIndexBufferEntryNVX, ObjectTablePipelineEntryNVX, ObjectTablePushConstantEntryNVX, ObjectTableVertexBufferEntryNVX, unregisterObjectsNVX

Constructors

ObjectEntryTypeNVX Int32

Bundled Patterns

pattern OBJECT_ENTRY_TYPE_DESCRIPTOR_SET_NVX :: ObjectEntryTypeNVX
OBJECT_ENTRY_TYPE_DESCRIPTOR_SET_NVX specifies a DescriptorSet resource entry that is registered via ObjectTableDescriptorSetEntryNVX.

pattern OBJECT_ENTRY_TYPE_PIPELINE_NVX :: ObjectEntryTypeNVX
OBJECT_ENTRY_TYPE_PIPELINE_NVX specifies a Pipeline resource entry that is registered via ObjectTablePipelineEntryNVX.

pattern OBJECT_ENTRY_TYPE_INDEX_BUFFER_NVX :: ObjectEntryTypeNVX
OBJECT_ENTRY_TYPE_INDEX_BUFFER_NVX specifies a Buffer resource entry that is registered via ObjectTableIndexBufferEntryNVX.

pattern OBJECT_ENTRY_TYPE_VERTEX_BUFFER_NVX :: ObjectEntryTypeNVX
OBJECT_ENTRY_TYPE_VERTEX_BUFFER_NVX specifies a Buffer resource entry that is registered via ObjectTableVertexBufferEntryNVX.

pattern OBJECT_ENTRY_TYPE_PUSH_CONSTANT_NVX :: ObjectEntryTypeNVX
OBJECT_ENTRY_TYPE_PUSH_CONSTANT_NVX specifies the resource entry is registered via ObjectTablePushConstantEntryNVX.

type NVX_DEVICE_GENERATED_COMMANDS_SPEC_VERSION = 3

pattern NVX_DEVICE_GENERATED_COMMANDS_SPEC_VERSION :: forall a. Integral a => a

type NVX_DEVICE_GENERATED_COMMANDS_EXTENSION_NAME = "VK_NVX_device_generated_commands"

pattern NVX_DEVICE_GENERATED_COMMANDS_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype ObjectTableNVX

VkObjectTableNVX - Opaque handle to an object table

See Also

CmdProcessCommandsInfoNVX, CmdReserveSpaceForCommandsInfoNVX, createObjectTableNVX, destroyObjectTableNVX, registerObjectsNVX, unregisterObjectsNVX

Constructors

ObjectTableNVX Word64

newtype IndirectCommandsLayoutNVX

VkIndirectCommandsLayoutNVX - Opaque handle to an indirect commands layout object

See Also

CmdProcessCommandsInfoNVX, CmdReserveSpaceForCommandsInfoNVX, createIndirectCommandsLayoutNVX, destroyIndirectCommandsLayoutNVX

Constructors

IndirectCommandsLayoutNVX Word64