Graphics.Vulkan.Extensions.VK_NVX_image_view_handle

Documentation

withCStructImageViewHandleInfoNVX :: ImageViewHandleInfoNVX -> (VkImageViewHandleInfoNVX -> IO a) -> IO a

A function to temporarily allocate memory for a VkImageViewHandleInfoNVX and marshal a ImageViewHandleInfoNVX into it. The VkImageViewHandleInfoNVX is only valid inside the provided computation and must not be returned out of it.

fromCStructImageViewHandleInfoNVX :: VkImageViewHandleInfoNVX -> IO ImageViewHandleInfoNVX

A function to read a VkImageViewHandleInfoNVX and all additional structures in the pointer chain into a ImageViewHandleInfoNVX.

data ImageViewHandleInfoNVX

VkImageViewHandleInfoNVX - Structure specifying the image view for handle queries

Valid Usage

descriptorType must be VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE, VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, or VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER

sampler must be a valid VkSampler if descriptorType is VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER

If descriptorType is VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE or VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, the image that imageView was created from must have been created with the VK_IMAGE_USAGE_SAMPLED_BIT usage bit set

If descriptorType is VK_DESCRIPTOR_TYPE_STORAGE_IMAGE, the image that imageView was created from must have been created with the VK_IMAGE_USAGE_STORAGE_BIT usage bit set

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMAGE_VIEW_HANDLE_INFO_NVX

pNext must be NULL

imageView must be a valid VkImageView handle

descriptorType must be a valid VkDescriptorType value

If sampler is not VK_NULL_HANDLE, sampler must be a valid VkSampler handle

Both of imageView, and sampler that are valid handles must have been created, allocated, or retrieved from the same VkDevice

See Also

VkDescriptorType, VkImageView, VkSampler, VkStructureType, vkGetImageViewHandleNVX

Constructors

ImageViewHandleInfoNVX

Fields

next :: Maybe SomeVkStruct

imageView :: ImageView

descriptorType :: DescriptorType

sampler :: Sampler

getImageViewHandleNVX :: Device -> ImageViewHandleInfoNVX -> IO Word32

vkGetImageViewHandleNVX - Get the handle for an image view for a specific descriptor type

Parameters

device is the logical device that owns the image view.

pInfo describes the image view to query and type of handle.

Valid Usage (Implicit)

See Also

VkDevice, VkImageViewHandleInfoNVX

pattern NVX_IMAGE_VIEW_HANDLE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NVX_IMAGE_VIEW_HANDLE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_IMAGE_VIEW_HANDLE_INFO_NVX :: a ~ StructureType => a