Graphics.Vulkan.Extensions.VK_NVX_multiview_per_view_attributes

Documentation

withCStructPhysicalDeviceMultiviewPerViewAttributesPropertiesNVX :: PhysicalDeviceMultiviewPerViewAttributesPropertiesNVX -> (VkPhysicalDeviceMultiviewPerViewAttributesPropertiesNVX -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceMultiviewPerViewAttributesPropertiesNVX and marshal a PhysicalDeviceMultiviewPerViewAttributesPropertiesNVX into it. The VkPhysicalDeviceMultiviewPerViewAttributesPropertiesNVX is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceMultiviewPerViewAttributesPropertiesNVX :: VkPhysicalDeviceMultiviewPerViewAttributesPropertiesNVX -> IO PhysicalDeviceMultiviewPerViewAttributesPropertiesNVX

A function to read a VkPhysicalDeviceMultiviewPerViewAttributesPropertiesNVX and all additional structures in the pointer chain into a PhysicalDeviceMultiviewPerViewAttributesPropertiesNVX.

data PhysicalDeviceMultiviewPerViewAttributesPropertiesNVX

VkPhysicalDeviceMultiviewPerViewAttributesPropertiesNVX - Structure describing multiview limits that can be supported by an implementation

Members

The members of the VkPhysicalDeviceMultiviewPerViewAttributesPropertiesNVX structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceMultiviewPerViewAttributesPropertiesNVX structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceMultiviewPerViewAttributesPropertiesNVX

Fields

next :: Maybe SomeVkStruct

perViewPositionAllComponents :: Bool

pattern NVX_MULTIVIEW_PER_VIEW_ATTRIBUTES_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NVX_MULTIVIEW_PER_VIEW_ATTRIBUTES_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_MULTIVIEW_PER_VIEW_ATTRIBUTES_PROPERTIES_NVX :: a ~ StructureType => a

pattern SUBPASS_DESCRIPTION_PER_VIEW_ATTRIBUTES_BIT_NVX :: a ~ SubpassDescriptionFlagBits => a

pattern SUBPASS_DESCRIPTION_PER_VIEW_POSITION_X_ONLY_BIT_NVX :: a ~ SubpassDescriptionFlagBits => a