Graphics.Vulkan.Extensions.VK_NV_compute_shader_derivatives

Documentation

withCStructPhysicalDeviceComputeShaderDerivativesFeaturesNV :: PhysicalDeviceComputeShaderDerivativesFeaturesNV -> (VkPhysicalDeviceComputeShaderDerivativesFeaturesNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceComputeShaderDerivativesFeaturesNV and marshal a PhysicalDeviceComputeShaderDerivativesFeaturesNV into it. The VkPhysicalDeviceComputeShaderDerivativesFeaturesNV is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceComputeShaderDerivativesFeaturesNV :: VkPhysicalDeviceComputeShaderDerivativesFeaturesNV -> IO PhysicalDeviceComputeShaderDerivativesFeaturesNV

A function to read a VkPhysicalDeviceComputeShaderDerivativesFeaturesNV and all additional structures in the pointer chain into a PhysicalDeviceComputeShaderDerivativesFeaturesNV.

data PhysicalDeviceComputeShaderDerivativesFeaturesNV

VkPhysicalDeviceComputeShaderDerivativesFeaturesNV - Structure describing compute shader derivative features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceComputeShaderDerivativesFeaturesNV structure describe the following features:

Description

See Compute Shader Derivatives for more information.

If the VkPhysicalDeviceComputeShaderDerivativesFeaturesNV structure is included in the pNext chain of VkPhysicalDeviceFeatures2KHR, it is filled with values indicating whether each feature is supported. VkPhysicalDeviceComputeShaderDerivativesFeaturesNV can also be used in the pNext chain of VkDeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceComputeShaderDerivativesFeaturesNV

Fields

next :: Maybe SomeVkStruct

computeDerivativeGroupQuads :: Bool

computeDerivativeGroupLinear :: Bool

pattern NV_COMPUTE_SHADER_DERIVATIVES_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NV_COMPUTE_SHADER_DERIVATIVES_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_COMPUTE_SHADER_DERIVATIVES_FEATURES_NV :: a ~ StructureType => a