Graphics.Vulkan.Extensions.VK_NV_cooperative_matrix

Documentation

type ComponentTypeNV = VkComponentTypeNV

VkComponentTypeNV - Specify SPIR-V cooperative matrix component type

See Also

VkCooperativeMatrixPropertiesNV

pattern COMPONENT_TYPE_FLOAT16_NV :: a ~ ComponentTypeNV => a

VK_COMPONENT_TYPE_FLOAT16_NV corresponds to SPIR-V OpTypeFloat 16.

pattern COMPONENT_TYPE_FLOAT32_NV :: a ~ ComponentTypeNV => a

VK_COMPONENT_TYPE_FLOAT32_NV corresponds to SPIR-V OpTypeFloat 32.

pattern COMPONENT_TYPE_FLOAT64_NV :: a ~ ComponentTypeNV => a

VK_COMPONENT_TYPE_FLOAT64_NV corresponds to SPIR-V OpTypeFloat 64.

pattern COMPONENT_TYPE_SINT8_NV :: a ~ ComponentTypeNV => a

VK_COMPONENT_TYPE_SINT8_NV corresponds to SPIR-V OpTypeInt 8 1.

pattern COMPONENT_TYPE_SINT16_NV :: a ~ ComponentTypeNV => a

VK_COMPONENT_TYPE_SINT16_NV corresponds to SPIR-V OpTypeInt 16 1.

pattern COMPONENT_TYPE_SINT32_NV :: a ~ ComponentTypeNV => a

VK_COMPONENT_TYPE_SINT32_NV corresponds to SPIR-V OpTypeInt 32 1.

pattern COMPONENT_TYPE_SINT64_NV :: a ~ ComponentTypeNV => a

VK_COMPONENT_TYPE_SINT64_NV corresponds to SPIR-V OpTypeInt 64 1.

pattern COMPONENT_TYPE_UINT8_NV :: a ~ ComponentTypeNV => a

VK_COMPONENT_TYPE_UINT8_NV corresponds to SPIR-V OpTypeInt 8 0.

pattern COMPONENT_TYPE_UINT16_NV :: a ~ ComponentTypeNV => a

VK_COMPONENT_TYPE_UINT16_NV corresponds to SPIR-V OpTypeInt 16 0.

pattern COMPONENT_TYPE_UINT32_NV :: a ~ ComponentTypeNV => a

VK_COMPONENT_TYPE_UINT32_NV corresponds to SPIR-V OpTypeInt 32 0.

pattern COMPONENT_TYPE_UINT64_NV :: a ~ ComponentTypeNV => a

VK_COMPONENT_TYPE_UINT64_NV corresponds to SPIR-V OpTypeInt 64 0.

withCStructCooperativeMatrixPropertiesNV :: CooperativeMatrixPropertiesNV -> (VkCooperativeMatrixPropertiesNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkCooperativeMatrixPropertiesNV and marshal a CooperativeMatrixPropertiesNV into it. The VkCooperativeMatrixPropertiesNV is only valid inside the provided computation and must not be returned out of it.

fromCStructCooperativeMatrixPropertiesNV :: VkCooperativeMatrixPropertiesNV -> IO CooperativeMatrixPropertiesNV

A function to read a VkCooperativeMatrixPropertiesNV and all additional structures in the pointer chain into a CooperativeMatrixPropertiesNV.

data CooperativeMatrixPropertiesNV

VkCooperativeMatrixPropertiesNV - Structure specifying cooperative matrix properties

Description

If some types are preferred over other types (e.g. for performance), they should appear earlier in the list enumerated by vkGetPhysicalDeviceCooperativeMatrixPropertiesNV.

At least one entry in the list must have power of two values for all of MSize, KSize, and NSize.

Valid Usage (Implicit)

See Also

VkComponentTypeNV, VkScopeNV, VkStructureType, vkGetPhysicalDeviceCooperativeMatrixPropertiesNV

Constructors

CooperativeMatrixPropertiesNV

Fields

next :: Maybe SomeVkStruct

mSize :: Word32

nSize :: Word32

kSize :: Word32

aType :: ComponentTypeNV

bType :: ComponentTypeNV

cType :: ComponentTypeNV

dType :: ComponentTypeNV

scope :: ScopeNV

withCStructPhysicalDeviceCooperativeMatrixFeaturesNV :: PhysicalDeviceCooperativeMatrixFeaturesNV -> (VkPhysicalDeviceCooperativeMatrixFeaturesNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceCooperativeMatrixFeaturesNV and marshal a PhysicalDeviceCooperativeMatrixFeaturesNV into it. The VkPhysicalDeviceCooperativeMatrixFeaturesNV is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceCooperativeMatrixFeaturesNV :: VkPhysicalDeviceCooperativeMatrixFeaturesNV -> IO PhysicalDeviceCooperativeMatrixFeaturesNV

A function to read a VkPhysicalDeviceCooperativeMatrixFeaturesNV and all additional structures in the pointer chain into a PhysicalDeviceCooperativeMatrixFeaturesNV.

data PhysicalDeviceCooperativeMatrixFeaturesNV

VkPhysicalDeviceCooperativeMatrixFeaturesNV - Structure describing cooperative matrix features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceCooperativeMatrixFeaturesNV structure describe the following features:

Description

If the VkPhysicalDeviceCooperativeMatrixFeaturesNV structure is included in the pNext chain of VkPhysicalDeviceFeatures2KHR, it is filled with values indicating whether the feature is supported. VkPhysicalDeviceCooperativeMatrixFeaturesNV can also be used in the pNext chain of VkDeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceCooperativeMatrixFeaturesNV

Fields

next :: Maybe SomeVkStruct

cooperativeMatrix :: Bool

cooperativeMatrixRobustBufferAccess :: Bool

withCStructPhysicalDeviceCooperativeMatrixPropertiesNV :: PhysicalDeviceCooperativeMatrixPropertiesNV -> (VkPhysicalDeviceCooperativeMatrixPropertiesNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceCooperativeMatrixPropertiesNV and marshal a PhysicalDeviceCooperativeMatrixPropertiesNV into it. The VkPhysicalDeviceCooperativeMatrixPropertiesNV is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceCooperativeMatrixPropertiesNV :: VkPhysicalDeviceCooperativeMatrixPropertiesNV -> IO PhysicalDeviceCooperativeMatrixPropertiesNV

A function to read a VkPhysicalDeviceCooperativeMatrixPropertiesNV and all additional structures in the pointer chain into a PhysicalDeviceCooperativeMatrixPropertiesNV.

data PhysicalDeviceCooperativeMatrixPropertiesNV

VkPhysicalDeviceCooperativeMatrixPropertiesNV - Structure describing cooperative matrix properties supported by an implementation

Members

The members of the VkPhysicalDeviceCooperativeMatrixPropertiesNV structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceCooperativeMatrixPropertiesNV structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkShaderStageFlags, VkStructureType

Constructors

PhysicalDeviceCooperativeMatrixPropertiesNV

Fields

next :: Maybe SomeVkStruct

cooperativeMatrixSupportedStages :: ShaderStageFlags

type ScopeNV = VkScopeNV

VkScopeNV - Specify SPIR-V scope

Description

All enum values match the corresponding SPIR-V value.

See Also

VkCooperativeMatrixPropertiesNV

pattern SCOPE_DEVICE_NV :: a ~ ScopeNV => a

VK_SCOPE_DEVICE_NV corresponds to SPIR-V Device scope.

pattern SCOPE_WORKGROUP_NV :: a ~ ScopeNV => a

VK_SCOPE_WORKGROUP_NV corresponds to SPIR-V Workgroup scope.

pattern SCOPE_SUBGROUP_NV :: a ~ ScopeNV => a

VK_SCOPE_SUBGROUP_NV corresponds to SPIR-V Subgroup scope.

pattern SCOPE_QUEUE_FAMILY_NV :: a ~ ScopeNV => a

VK_SCOPE_QUEUE_FAMILY_NV corresponds to SPIR-V QueueFamilyKHR scope.

getNumPhysicalDeviceCooperativeMatrixPropertiesNV :: PhysicalDevice -> IO (VkResult, Word32)

vkGetPhysicalDeviceCooperativeMatrixPropertiesNV - Returns properties describing what cooperative matrix types are supported

Parameters

physicalDevice is the physical device.

pPropertyCount is a pointer to an integer related to the number of cooperative matrix properties available or queried.

pProperties is either NULL or a pointer to an array of VkCooperativeMatrixPropertiesNV structures.

Description

If pProperties is NULL, then the number of cooperative matrix properties available is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If pPropertyCount is less than the number of cooperative matrix properties available, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of cooperative matrix properties available, VK_INCOMPLETE will be returned instead of VK_SUCCESS, to indicate that not all the available cooperative matrix properties were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkCooperativeMatrixPropertiesNV structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkCooperativeMatrixPropertiesNV, VkPhysicalDevice

getPhysicalDeviceCooperativeMatrixPropertiesNV :: PhysicalDevice -> Word32 -> IO (VkResult, Vector CooperativeMatrixPropertiesNV)

vkGetPhysicalDeviceCooperativeMatrixPropertiesNV - Returns properties describing what cooperative matrix types are supported

Parameters

physicalDevice is the physical device.

pPropertyCount is a pointer to an integer related to the number of cooperative matrix properties available or queried.

pProperties is either NULL or a pointer to an array of VkCooperativeMatrixPropertiesNV structures.

Description

If pProperties is NULL, then the number of cooperative matrix properties available is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If pPropertyCount is less than the number of cooperative matrix properties available, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of cooperative matrix properties available, VK_INCOMPLETE will be returned instead of VK_SUCCESS, to indicate that not all the available cooperative matrix properties were returned.

Valid Usage (Implicit)

physicalDevice must be a valid VkPhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount VkCooperativeMatrixPropertiesNV structures

Return Codes

Success

- VK_SUCCESS

VK_INCOMPLETE

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkCooperativeMatrixPropertiesNV, VkPhysicalDevice

getAllPhysicalDeviceCooperativeMatrixPropertiesNV :: PhysicalDevice -> IO (Vector CooperativeMatrixPropertiesNV)

Returns all the values available from getPhysicalDeviceCooperativeMatrixPropertiesNV.

pattern NV_COOPERATIVE_MATRIX_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NV_COOPERATIVE_MATRIX_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_COOPERATIVE_MATRIX_FEATURES_NV :: a ~ StructureType => a

pattern STRUCTURE_TYPE_COOPERATIVE_MATRIX_PROPERTIES_NV :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_COOPERATIVE_MATRIX_PROPERTIES_NV :: a ~ StructureType => a