Graphics.Vulkan.Extensions.VK_NV_dedicated_allocation

Documentation

withCStructDedicatedAllocationBufferCreateInfoNV :: DedicatedAllocationBufferCreateInfoNV -> (VkDedicatedAllocationBufferCreateInfoNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkDedicatedAllocationBufferCreateInfoNV and marshal a DedicatedAllocationBufferCreateInfoNV into it. The VkDedicatedAllocationBufferCreateInfoNV is only valid inside the provided computation and must not be returned out of it.

fromCStructDedicatedAllocationBufferCreateInfoNV :: VkDedicatedAllocationBufferCreateInfoNV -> IO DedicatedAllocationBufferCreateInfoNV

A function to read a VkDedicatedAllocationBufferCreateInfoNV and all additional structures in the pointer chain into a DedicatedAllocationBufferCreateInfoNV.

data DedicatedAllocationBufferCreateInfoNV

VkDedicatedAllocationBufferCreateInfoNV - Specify that a buffer is bound to a dedicated memory resource

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

DedicatedAllocationBufferCreateInfoNV

Fields

next :: Maybe SomeVkStruct

dedicatedAllocation :: Bool

withCStructDedicatedAllocationImageCreateInfoNV :: DedicatedAllocationImageCreateInfoNV -> (VkDedicatedAllocationImageCreateInfoNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkDedicatedAllocationImageCreateInfoNV and marshal a DedicatedAllocationImageCreateInfoNV into it. The VkDedicatedAllocationImageCreateInfoNV is only valid inside the provided computation and must not be returned out of it.

fromCStructDedicatedAllocationImageCreateInfoNV :: VkDedicatedAllocationImageCreateInfoNV -> IO DedicatedAllocationImageCreateInfoNV

A function to read a VkDedicatedAllocationImageCreateInfoNV and all additional structures in the pointer chain into a DedicatedAllocationImageCreateInfoNV.

data DedicatedAllocationImageCreateInfoNV

VkDedicatedAllocationImageCreateInfoNV - Specify that an image is bound to a dedicated memory resource

Description

Note

Using a dedicated allocation for color and depth/stencil attachments or other large images may improve performance on some devices.

Valid Usage

If dedicatedAllocation is VK_TRUE, VkImageCreateInfo::flags must not include VK_IMAGE_CREATE_SPARSE_BINDING_BIT, VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT, or VK_IMAGE_CREATE_SPARSE_ALIASED_BIT

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DEDICATED_ALLOCATION_IMAGE_CREATE_INFO_NV

See Also

VkBool32, VkStructureType

Constructors

DedicatedAllocationImageCreateInfoNV

Fields

next :: Maybe SomeVkStruct

dedicatedAllocation :: Bool

withCStructDedicatedAllocationMemoryAllocateInfoNV :: DedicatedAllocationMemoryAllocateInfoNV -> (VkDedicatedAllocationMemoryAllocateInfoNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkDedicatedAllocationMemoryAllocateInfoNV and marshal a DedicatedAllocationMemoryAllocateInfoNV into it. The VkDedicatedAllocationMemoryAllocateInfoNV is only valid inside the provided computation and must not be returned out of it.

fromCStructDedicatedAllocationMemoryAllocateInfoNV :: VkDedicatedAllocationMemoryAllocateInfoNV -> IO DedicatedAllocationMemoryAllocateInfoNV

A function to read a VkDedicatedAllocationMemoryAllocateInfoNV and all additional structures in the pointer chain into a DedicatedAllocationMemoryAllocateInfoNV.

data DedicatedAllocationMemoryAllocateInfoNV

VkDedicatedAllocationMemoryAllocateInfoNV - Specify a dedicated memory allocation resource

Valid Usage

At least one of image and buffer must be VK_NULL_HANDLE

If image is not VK_NULL_HANDLE, the image must have been created with VkDedicatedAllocationImageCreateInfoNV::dedicatedAllocation equal to VK_TRUE

If buffer is not VK_NULL_HANDLE, the buffer must have been created with VkDedicatedAllocationBufferCreateInfoNV::dedicatedAllocation equal to VK_TRUE

If image is not VK_NULL_HANDLE, VkMemoryAllocateInfo::allocationSize must equal the VkMemoryRequirements::size of the image

If buffer is not VK_NULL_HANDLE, VkMemoryAllocateInfo::allocationSize must equal the VkMemoryRequirements::size of the buffer

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_DEDICATED_ALLOCATION_MEMORY_ALLOCATE_INFO_NV

If image is not VK_NULL_HANDLE, image must be a valid VkImage handle

If buffer is not VK_NULL_HANDLE, buffer must be a valid VkBuffer handle

Both of buffer, and image that are valid handles must have been created, allocated, or retrieved from the same VkDevice

See Also

VkBuffer, VkImage, VkStructureType

Constructors

DedicatedAllocationMemoryAllocateInfoNV

Fields

next :: Maybe SomeVkStruct

image :: Image

buffer :: Buffer

pattern NV_DEDICATED_ALLOCATION_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NV_DEDICATED_ALLOCATION_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_DEDICATED_ALLOCATION_IMAGE_CREATE_INFO_NV :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEDICATED_ALLOCATION_BUFFER_CREATE_INFO_NV :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEDICATED_ALLOCATION_MEMORY_ALLOCATE_INFO_NV :: a ~ StructureType => a