Graphics.Vulkan.Extensions.VK_NV_external_memory

Documentation

withCStructExportMemoryAllocateInfoNV :: ExportMemoryAllocateInfoNV -> (VkExportMemoryAllocateInfoNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkExportMemoryAllocateInfoNV and marshal a ExportMemoryAllocateInfoNV into it. The VkExportMemoryAllocateInfoNV is only valid inside the provided computation and must not be returned out of it.

fromCStructExportMemoryAllocateInfoNV :: VkExportMemoryAllocateInfoNV -> IO ExportMemoryAllocateInfoNV

A function to read a VkExportMemoryAllocateInfoNV and all additional structures in the pointer chain into a ExportMemoryAllocateInfoNV.

data ExportMemoryAllocateInfoNV

VkExportMemoryAllocateInfoNV - Specify memory handle types that may be exported

Valid Usage (Implicit)

See Also

VkExternalMemoryHandleTypeFlagsNV, VkStructureType

Constructors

ExportMemoryAllocateInfoNV

Fields

next :: Maybe SomeVkStruct

handleTypes :: ExternalMemoryHandleTypeFlagsNV

withCStructExternalMemoryImageCreateInfoNV :: ExternalMemoryImageCreateInfoNV -> (VkExternalMemoryImageCreateInfoNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkExternalMemoryImageCreateInfoNV and marshal a ExternalMemoryImageCreateInfoNV into it. The VkExternalMemoryImageCreateInfoNV is only valid inside the provided computation and must not be returned out of it.

fromCStructExternalMemoryImageCreateInfoNV :: VkExternalMemoryImageCreateInfoNV -> IO ExternalMemoryImageCreateInfoNV

A function to read a VkExternalMemoryImageCreateInfoNV and all additional structures in the pointer chain into a ExternalMemoryImageCreateInfoNV.

data ExternalMemoryImageCreateInfoNV

VkExternalMemoryImageCreateInfoNV - Specify that an image may be backed by external memory

Valid Usage (Implicit)

See Also

VkExternalMemoryHandleTypeFlagsNV, VkStructureType

Constructors

ExternalMemoryImageCreateInfoNV

Fields

next :: Maybe SomeVkStruct

handleTypes :: ExternalMemoryHandleTypeFlagsNV

pattern NV_EXTERNAL_MEMORY_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NV_EXTERNAL_MEMORY_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_EXTERNAL_MEMORY_IMAGE_CREATE_INFO_NV :: a ~ StructureType => a

pattern STRUCTURE_TYPE_EXPORT_MEMORY_ALLOCATE_INFO_NV :: a ~ StructureType => a