Graphics.Vulkan.Extensions.VK_NV_external_memory_capabilities

Documentation

withCStructExternalImageFormatPropertiesNV :: ExternalImageFormatPropertiesNV -> (VkExternalImageFormatPropertiesNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkExternalImageFormatPropertiesNV and marshal a ExternalImageFormatPropertiesNV into it. The VkExternalImageFormatPropertiesNV is only valid inside the provided computation and must not be returned out of it.

fromCStructExternalImageFormatPropertiesNV :: VkExternalImageFormatPropertiesNV -> IO ExternalImageFormatPropertiesNV

A function to read a VkExternalImageFormatPropertiesNV and all additional structures in the pointer chain into a ExternalImageFormatPropertiesNV.

data ExternalImageFormatPropertiesNV

VkExternalImageFormatPropertiesNV - Structure specifying external image format properties

See Also

VkExternalMemoryFeatureFlagsNV, VkExternalMemoryHandleTypeFlagsNV, VkImageFormatProperties, vkGetPhysicalDeviceExternalImageFormatPropertiesNV

Constructors

ExternalImageFormatPropertiesNV

Fields

imageFormatProperties :: ImageFormatProperties

externalMemoryFeatures :: ExternalMemoryFeatureFlagsNV

exportFromImportedHandleTypes :: ExternalMemoryHandleTypeFlagsNV

compatibleHandleTypes :: ExternalMemoryHandleTypeFlagsNV

type ExternalMemoryFeatureFlagBitsNV = VkExternalMemoryFeatureFlagBitsNV

VkExternalMemoryFeatureFlagBitsNV - Bitmask specifying external memory features

See Also

VkExternalImageFormatPropertiesNV, VkExternalMemoryFeatureFlagsNV, vkGetPhysicalDeviceExternalImageFormatPropertiesNV

pattern EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT_NV :: a ~ ExternalMemoryFeatureFlagBitsNV => a

VK_EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT_NV specifies that external memory of the specified type must be created as a dedicated allocation when used in the manner specified.

pattern EXTERNAL_MEMORY_FEATURE_EXPORTABLE_BIT_NV :: a ~ ExternalMemoryFeatureFlagBitsNV => a

VK_EXTERNAL_MEMORY_FEATURE_EXPORTABLE_BIT_NV specifies that the implementation supports exporting handles of the specified type.

pattern EXTERNAL_MEMORY_FEATURE_IMPORTABLE_BIT_NV :: a ~ ExternalMemoryFeatureFlagBitsNV => a

VK_EXTERNAL_MEMORY_FEATURE_IMPORTABLE_BIT_NV specifies that the implementation supports importing handles of the specified type.

type ExternalMemoryFeatureFlagsNV = ExternalMemoryFeatureFlagBitsNV

VkExternalMemoryFeatureFlagsNV - Bitmask of VkExternalMemoryFeatureFlagBitsNV

Description

VkExternalMemoryFeatureFlagsNV is a bitmask type for setting a mask of zero or more VkExternalMemoryFeatureFlagBitsNV.

See Also

VkExternalImageFormatPropertiesNV, VkExternalMemoryFeatureFlagBitsNV

type ExternalMemoryHandleTypeFlagBitsNV = VkExternalMemoryHandleTypeFlagBitsNV

VkExternalMemoryHandleTypeFlagBitsNV - Bitmask specifying external memory handle types

See Also

VkExternalMemoryHandleTypeFlagsNV

pattern EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT_NV :: a ~ ExternalMemoryHandleTypeFlagBitsNV => a

VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT_NV specifies a handle to memory returned by vkGetMemoryWin32HandleNV, or one duplicated from such a handle using DuplicateHandle().

pattern EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_NV :: a ~ ExternalMemoryHandleTypeFlagBitsNV => a

VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_NV specifies a handle to memory returned by vkGetMemoryWin32HandleNV.

pattern EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_BIT_NV :: a ~ ExternalMemoryHandleTypeFlagBitsNV => a

VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_BIT_NV specifies a valid NT handle to memory returned by IDXGIResource1::CreateSharedHandle, or a handle duplicated from such a handle using DuplicateHandle().

pattern EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_KMT_BIT_NV :: a ~ ExternalMemoryHandleTypeFlagBitsNV => a

VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_KMT_BIT_NV specifies a handle to memory returned by IDXGIResource::GetSharedHandle().

type ExternalMemoryHandleTypeFlagsNV = ExternalMemoryHandleTypeFlagBitsNV

VkExternalMemoryHandleTypeFlagsNV - Bitmask of VkExternalMemoryHandleTypeFlagBitsNV

Description

VkExternalMemoryHandleTypeFlagsNV is a bitmask type for setting a mask of zero or more VkExternalMemoryHandleTypeFlagBitsNV.

See Also

VkExportMemoryAllocateInfoNV, VkExternalImageFormatPropertiesNV, VkExternalMemoryHandleTypeFlagBitsNV, VkExternalMemoryImageCreateInfoNV, VkImportMemoryWin32HandleInfoNV, vkGetMemoryWin32HandleNV, vkGetPhysicalDeviceExternalImageFormatPropertiesNV

getPhysicalDeviceExternalImageFormatPropertiesNV :: PhysicalDevice -> Format -> ImageType -> ImageTiling -> ImageUsageFlags -> ImageCreateFlags -> ExternalMemoryHandleTypeFlagsNV -> IO ExternalImageFormatPropertiesNV

vkGetPhysicalDeviceExternalImageFormatPropertiesNV - determine image capabilities compatible with external memory handle types

Parameters

physicalDevice is the physical device from which to query the image capabilities

format is the image format, corresponding to VkImageCreateInfo::format.

type is the image type, corresponding to VkImageCreateInfo::imageType.

tiling is the image tiling, corresponding to VkImageCreateInfo::tiling.

usage is the intended usage of the image, corresponding to VkImageCreateInfo::usage.

flags is a bitmask describing additional parameters of the image, corresponding to VkImageCreateInfo::flags.

externalHandleType is either one of the bits from VkExternalMemoryHandleTypeFlagBitsNV, or 0.

pExternalImageFormatProperties points to an instance of the VkExternalImageFormatPropertiesNV structure in which capabilities are returned.

Description

If externalHandleType is 0, pExternalImageFormatProperties::imageFormatProperties will return the same values as a call to vkGetPhysicalDeviceImageFormatProperties, and the other members of pExternalImageFormatProperties will all be 0. Otherwise, they are filled in as described for VkExternalImageFormatPropertiesNV.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_FORMAT_NOT_SUPPORTED

See Also

VkExternalImageFormatPropertiesNV, VkExternalMemoryHandleTypeFlagsNV, VkFormat, VkImageCreateFlags, VkImageTiling, VkImageType, VkImageUsageFlags, VkPhysicalDevice

pattern NV_EXTERNAL_MEMORY_CAPABILITIES_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NV_EXTERNAL_MEMORY_CAPABILITIES_SPEC_VERSION :: Integral a => a