Graphics.Vulkan.Extensions.VK_NV_external_memory_win32

Documentation

withCStructExportMemoryWin32HandleInfoNV :: ExportMemoryWin32HandleInfoNV -> (VkExportMemoryWin32HandleInfoNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkExportMemoryWin32HandleInfoNV and marshal a ExportMemoryWin32HandleInfoNV into it. The VkExportMemoryWin32HandleInfoNV is only valid inside the provided computation and must not be returned out of it.

fromCStructExportMemoryWin32HandleInfoNV :: VkExportMemoryWin32HandleInfoNV -> IO ExportMemoryWin32HandleInfoNV

A function to read a VkExportMemoryWin32HandleInfoNV and all additional structures in the pointer chain into a ExportMemoryWin32HandleInfoNV.

data ExportMemoryWin32HandleInfoNV

VkExportMemoryWin32HandleInfoNV - specify security attributes and access rights for Win32 memory handles

Description

If this structure is not present, or if pAttributes is set to NULL, default security descriptor values will be used, and child processes created by the application will not inherit the handle, as described in the MSDN documentation for “Synchronization Object Security and Access Rights”[1]. Further, if the structure is not present, the access rights will be

DXGI_SHARED_RESOURCE_READ | DXGI_SHARED_RESOURCE_WRITE

1

https://msdn.microsoft.com/en-us/library/windows/desktop/ms686670.aspx

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_EXPORT_MEMORY_WIN32_HANDLE_INFO_NV

If pAttributes is not NULL, pAttributes must be a valid pointer to a valid SECURITY_ATTRIBUTES value

See Also

VkStructureType

Constructors

ExportMemoryWin32HandleInfoNV

Fields

next :: Maybe SomeVkStruct

attributes :: Ptr SECURITY_ATTRIBUTES

dwAccess :: DWORD

withCStructImportMemoryWin32HandleInfoNV :: ImportMemoryWin32HandleInfoNV -> (VkImportMemoryWin32HandleInfoNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkImportMemoryWin32HandleInfoNV and marshal a ImportMemoryWin32HandleInfoNV into it. The VkImportMemoryWin32HandleInfoNV is only valid inside the provided computation and must not be returned out of it.

fromCStructImportMemoryWin32HandleInfoNV :: VkImportMemoryWin32HandleInfoNV -> IO ImportMemoryWin32HandleInfoNV

A function to read a VkImportMemoryWin32HandleInfoNV and all additional structures in the pointer chain into a ImportMemoryWin32HandleInfoNV.

data ImportMemoryWin32HandleInfoNV

VkImportMemoryWin32HandleInfoNV - import Win32 memory created on the same physical device

Description

If handleType is 0, this structure is ignored by consumers of the VkMemoryAllocateInfo structure it is chained from.

Valid Usage (Implicit)

See Also

VkExternalMemoryHandleTypeFlagsNV, VkStructureType

Constructors

ImportMemoryWin32HandleInfoNV

Fields

next :: Maybe SomeVkStruct

handleType :: ExternalMemoryHandleTypeFlagsNV

handle :: HANDLE

getMemoryWin32HandleNV :: Device -> DeviceMemory -> ExternalMemoryHandleTypeFlagsNV -> IO HANDLE

vkGetMemoryWin32HandleNV - retrieve Win32 handle to a device memory object

Parameters

device is the logical device that owns the memory.

memory is the VkDeviceMemory object.

handleType is a bitmask of VkExternalMemoryHandleTypeFlagBitsNV containing a single bit specifying the type of handle requested.

handle points to a Windows HANDLE in which the handle is returned.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_TOO_MANY_OBJECTS

VK_ERROR_OUT_OF_HOST_MEMORY

See Also

VkDevice, VkDeviceMemory, VkExternalMemoryHandleTypeFlagsNV

pattern NV_EXTERNAL_MEMORY_WIN32_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NV_EXTERNAL_MEMORY_WIN32_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_IMPORT_MEMORY_WIN32_HANDLE_INFO_NV :: a ~ StructureType => a

pattern STRUCTURE_TYPE_EXPORT_MEMORY_WIN32_HANDLE_INFO_NV :: a ~ StructureType => a