Graphics.Vulkan.Extensions.VK_NV_fragment_shader_barycentric

Documentation

withCStructPhysicalDeviceFragmentShaderBarycentricFeaturesNV :: PhysicalDeviceFragmentShaderBarycentricFeaturesNV -> (VkPhysicalDeviceFragmentShaderBarycentricFeaturesNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceFragmentShaderBarycentricFeaturesNV and marshal a PhysicalDeviceFragmentShaderBarycentricFeaturesNV into it. The VkPhysicalDeviceFragmentShaderBarycentricFeaturesNV is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceFragmentShaderBarycentricFeaturesNV :: VkPhysicalDeviceFragmentShaderBarycentricFeaturesNV -> IO PhysicalDeviceFragmentShaderBarycentricFeaturesNV

A function to read a VkPhysicalDeviceFragmentShaderBarycentricFeaturesNV and all additional structures in the pointer chain into a PhysicalDeviceFragmentShaderBarycentricFeaturesNV.

data PhysicalDeviceFragmentShaderBarycentricFeaturesNV

VkPhysicalDeviceFragmentShaderBarycentricFeaturesNV - Structure describing barycentric support in fragment shaders that can be supported by an implementation

Members

The members of the VkPhysicalDeviceFragmentShaderBarycentricFeaturesNV structure describe the following features:

Description

See Barycentric Interpolation for more information.

If the VkPhysicalDeviceFragmentShaderBarycentricFeaturesNV structure is included in the pNext chain of VkPhysicalDeviceFeatures2KHR, it is filled with values indicating whether the feature is supported. VkPhysicalDeviceFragmentShaderBarycentricFeaturesNV can also be used in the pNext chain of VkDeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceFragmentShaderBarycentricFeaturesNV

Fields

next :: Maybe SomeVkStruct

fragmentShaderBarycentric :: Bool

pattern NV_FRAGMENT_SHADER_BARYCENTRIC_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NV_FRAGMENT_SHADER_BARYCENTRIC_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_FRAGMENT_SHADER_BARYCENTRIC_FEATURES_NV :: a ~ StructureType => a