Graphics.Vulkan.Extensions.VK_NV_framebuffer_mixed_samples

Documentation

type CoverageModulationModeNV = VkCoverageModulationModeNV

VkCoverageModulationModeNV - Specify the discard rectangle mode

See Also

VkPipelineCoverageModulationStateCreateInfoNV

pattern COVERAGE_MODULATION_MODE_NONE_NV :: a ~ CoverageModulationModeNV => a

VK_COVERAGE_MODULATION_MODE_NONE_NV specifies that no components are multiplied by the modulation factor.

pattern COVERAGE_MODULATION_MODE_RGB_NV :: a ~ CoverageModulationModeNV => a

VK_COVERAGE_MODULATION_MODE_RGB_NV specifies that the red, green, and blue components are multiplied by the modulation factor.

pattern COVERAGE_MODULATION_MODE_ALPHA_NV :: a ~ CoverageModulationModeNV => a

VK_COVERAGE_MODULATION_MODE_ALPHA_NV specifies that the alpha component is multiplied by the modulation factor.

pattern COVERAGE_MODULATION_MODE_RGBA_NV :: a ~ CoverageModulationModeNV => a

VK_COVERAGE_MODULATION_MODE_RGBA_NV specifies that all components are multiplied by the modulation factor.

type PipelineCoverageModulationStateCreateFlagsNV = VkPipelineCoverageModulationStateCreateFlagsNV

VkPipelineCoverageModulationStateCreateFlagsNV - Reserved for future use

Description

VkPipelineCoverageModulationStateCreateFlagsNV is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkPipelineCoverageModulationStateCreateInfoNV

withCStructPipelineCoverageModulationStateCreateInfoNV :: PipelineCoverageModulationStateCreateInfoNV -> (VkPipelineCoverageModulationStateCreateInfoNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkPipelineCoverageModulationStateCreateInfoNV and marshal a PipelineCoverageModulationStateCreateInfoNV into it. The VkPipelineCoverageModulationStateCreateInfoNV is only valid inside the provided computation and must not be returned out of it.

fromCStructPipelineCoverageModulationStateCreateInfoNV :: VkPipelineCoverageModulationStateCreateInfoNV -> IO PipelineCoverageModulationStateCreateInfoNV

A function to read a VkPipelineCoverageModulationStateCreateInfoNV and all additional structures in the pointer chain into a PipelineCoverageModulationStateCreateInfoNV.

data PipelineCoverageModulationStateCreateInfoNV

VkPipelineCoverageModulationStateCreateInfoNV - Structure specifying parameters controlling coverage modulation

Description

If coverageModulationTableEnable is VK_FALSE, then for each color sample the associated bits of the fragment’s coverage are counted and divided by the number of associated bits to produce a modulation factor R in the range (0,1] (a value of zero would have been killed due to a color coverage of 0). Specifically:

N = value of rasterizationSamples

M = value of VkAttachmentDescription::samples for any color attachments

R = popcount(associated coverage bits) / (N / M)

If coverageModulationTableEnable is VK_TRUE, the value R is computed using a programmable lookup table. The lookup table has N / M elements, and the element of the table is selected by:

R = pCoverageModulationTable[popcount(associated coverage bits)-1]

Note that the table does not have an entry for popcount(associated coverage bits) = 0, because such samples would have been killed.

The values of pCoverageModulationTable may be rounded to an implementation-dependent precision, which is at least as fine as 1 / N, and clamped to [0,1].

For each color attachment with a floating point or normalized color format, each fragment output color value is replicated to M values which can each be modulated (multiplied) by that color sample’s associated value of R. Which components are modulated is controlled by coverageModulationMode.

If this structure is not present, it is as if coverageModulationMode is VK_COVERAGE_MODULATION_MODE_NONE_NV.

Valid Usage

If coverageModulationTableEnable is VK_TRUE, coverageModulationTableCount must be equal to the number of rasterization samples divided by the number of color samples in the subpass.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PIPELINE_COVERAGE_MODULATION_STATE_CREATE_INFO_NV

flags must be 0

coverageModulationMode must be a valid VkCoverageModulationModeNV value

See Also

VkBool32, VkCoverageModulationModeNV, VkPipelineCoverageModulationStateCreateFlagsNV, VkStructureType

Constructors

PipelineCoverageModulationStateCreateInfoNV

Fields

next :: Maybe SomeVkStruct

flags :: PipelineCoverageModulationStateCreateFlagsNV

coverageModulationMode :: CoverageModulationModeNV

coverageModulationTableEnable :: Bool

coverageModulationTable :: Maybe (Vector CFloat)

pattern NV_FRAMEBUFFER_MIXED_SAMPLES_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NV_FRAMEBUFFER_MIXED_SAMPLES_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PIPELINE_COVERAGE_MODULATION_STATE_CREATE_INFO_NV :: a ~ StructureType => a