Graphics.Vulkan.Extensions.VK_NV_mesh_shader

Documentation

withCStructDrawMeshTasksIndirectCommandNV :: DrawMeshTasksIndirectCommandNV -> (VkDrawMeshTasksIndirectCommandNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkDrawMeshTasksIndirectCommandNV and marshal a DrawMeshTasksIndirectCommandNV into it. The VkDrawMeshTasksIndirectCommandNV is only valid inside the provided computation and must not be returned out of it.

fromCStructDrawMeshTasksIndirectCommandNV :: VkDrawMeshTasksIndirectCommandNV -> IO DrawMeshTasksIndirectCommandNV

A function to read a VkDrawMeshTasksIndirectCommandNV and all additional structures in the pointer chain into a DrawMeshTasksIndirectCommandNV.

data DrawMeshTasksIndirectCommandNV

VkDrawMeshTasksIndirectCommandNV - Structure specifying a mesh tasks draw indirect command

Description

The members of VkDrawMeshTasksIndirectCommandNV have the same meaning as the similarly named parameters of vkCmdDrawMeshTasksNV.

Valid Usage

See Also

vkCmdDrawMeshTasksIndirectNV

Constructors

DrawMeshTasksIndirectCommandNV

Fields

taskCount :: Word32

firstTask :: Word32

withCStructPhysicalDeviceMeshShaderFeaturesNV :: PhysicalDeviceMeshShaderFeaturesNV -> (VkPhysicalDeviceMeshShaderFeaturesNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceMeshShaderFeaturesNV and marshal a PhysicalDeviceMeshShaderFeaturesNV into it. The VkPhysicalDeviceMeshShaderFeaturesNV is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceMeshShaderFeaturesNV :: VkPhysicalDeviceMeshShaderFeaturesNV -> IO PhysicalDeviceMeshShaderFeaturesNV

A function to read a VkPhysicalDeviceMeshShaderFeaturesNV and all additional structures in the pointer chain into a PhysicalDeviceMeshShaderFeaturesNV.

data PhysicalDeviceMeshShaderFeaturesNV

VkPhysicalDeviceMeshShaderFeaturesNV - Structure describing mesh shading features that can be supported by an implementation

Description

If the VkPhysicalDeviceMeshShaderFeaturesNV structure is included in the pNext chain of VkPhysicalDeviceFeatures2, it is filled with a value indicating whether the feature is supported. VkPhysicalDeviceMeshShaderFeaturesNV can also be used in pNext chain of VkDeviceCreateInfo to enable the features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceMeshShaderFeaturesNV

Fields

next :: Maybe SomeVkStruct

taskShader :: Bool

meshShader :: Bool

withCStructPhysicalDeviceMeshShaderPropertiesNV :: PhysicalDeviceMeshShaderPropertiesNV -> (VkPhysicalDeviceMeshShaderPropertiesNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceMeshShaderPropertiesNV and marshal a PhysicalDeviceMeshShaderPropertiesNV into it. The VkPhysicalDeviceMeshShaderPropertiesNV is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceMeshShaderPropertiesNV :: VkPhysicalDeviceMeshShaderPropertiesNV -> IO PhysicalDeviceMeshShaderPropertiesNV

A function to read a VkPhysicalDeviceMeshShaderPropertiesNV and all additional structures in the pointer chain into a PhysicalDeviceMeshShaderPropertiesNV.

data PhysicalDeviceMeshShaderPropertiesNV

VkPhysicalDeviceMeshShaderPropertiesNV - Structure describing mesh shading properties

Members

The members of the VkPhysicalDeviceMeshShaderPropertiesNV structure describe the following implementation-dependent limits:

Description

If the VkPhysicalDeviceMeshShaderPropertiesNV structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkStructureType

Constructors

PhysicalDeviceMeshShaderPropertiesNV

Fields

next :: Maybe SomeVkStruct

maxDrawMeshTasksCount :: Word32

maxTaskWorkGroupInvocations :: Word32

maxTaskWorkGroupSize :: (Word32, Word32, Word32)

maxTaskTotalMemorySize :: Word32

maxTaskOutputCount :: Word32

maxMeshWorkGroupInvocations :: Word32

maxMeshWorkGroupSize :: (Word32, Word32, Word32)

maxMeshTotalMemorySize :: Word32

maxMeshOutputVertices :: Word32

maxMeshOutputPrimitives :: Word32

maxMeshMultiviewViewCount :: Word32

meshOutputPerVertexGranularity :: Word32

meshOutputPerPrimitiveGranularity :: Word32

cmdDrawMeshTasksIndirectCountNV :: CommandBuffer -> Buffer -> DeviceSize -> Buffer -> DeviceSize -> Word32 -> Word32 -> IO ()

vkCmdDrawMeshTasksIndirectCountNV - Perform an indirect mesh tasks draw with the draw count sourced from a buffer

Parameters

commandBuffer is the command buffer into which the command is recorded.

buffer is the buffer containing draw parameters.

offset is the byte offset into buffer where parameters begin.

countBuffer is the buffer containing the draw count.

countBufferOffset is the byte offset into countBuffer where the draw count begins.

maxDrawCount specifies the maximum number of draws that will be executed. The actual number of executed draw calls is the minimum of the count specified in countBuffer and maxDrawCount.

stride is the byte stride between successive sets of draw parameters.

Description

vkCmdDrawMeshTasksIndirectCountNV behaves similarly to vkCmdDrawMeshTasksIndirectNV except that the draw count is read by the device from a buffer during execution. The command will read an unsigned 32-bit integer from countBuffer located at countBufferOffset and use this as the draw count.

Valid Usage

If a VkImageView is sampled with VK_FILTER_LINEAR as a result of this command, then the image view’s format features must contain VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_LINEAR_BIT

If a VkImageView is accessed using atomic operations as a result of this command, then the image view’s format features must contain VK_FORMAT_FEATURE_STORAGE_IMAGE_ATOMIC_BIT

For each set n that is statically used by the VkPipeline bound to the pipeline bind point used by this command, a descriptor set must have been bound to n at the same pipeline bind point, with a VkPipelineLayout that is compatible for set n, with the VkPipelineLayout used to create the current VkPipeline, as described in ???

For each push constant that is statically used by the VkPipeline bound to the pipeline bind point used by this command, a push constant value must have been set for the same pipeline bind point, with a VkPipelineLayout that is compatible for push constants, with the VkPipelineLayout used to create the current VkPipeline, as described in ???

Descriptors in each bound descriptor set, specified via vkCmdBindDescriptorSets, must be valid if they are statically used by the VkPipeline bound to the pipeline bind point used by this command

A valid pipeline must be bound to the pipeline bind point used by this command

If the VkPipeline object bound to the pipeline bind point used by this command requires any dynamic state, that state must have been set for commandBuffer

If the VkPipeline object bound to the pipeline bind point used by this command accesses a VkSampler object that uses unnormalized coordinates, that sampler must not be used to sample from any VkImage with a VkImageView of the type VK_IMAGE_VIEW_TYPE_3D, VK_IMAGE_VIEW_TYPE_CUBE, VK_IMAGE_VIEW_TYPE_1D_ARRAY, VK_IMAGE_VIEW_TYPE_2D_ARRAY or VK_IMAGE_VIEW_TYPE_CUBE_ARRAY, in any shader stage

If the VkPipeline object bound to the pipeline bind point used by this command accesses a VkSampler object that uses unnormalized coordinates, that sampler must not be used with any of the SPIR-V OpImageSample* or OpImageSparseSample* instructions with ImplicitLod, Dref or Proj in their name, in any shader stage

If the VkPipeline object bound to the pipeline bind point used by this command accesses a VkSampler object that uses unnormalized coordinates, that sampler must not be used with any of the SPIR-V OpImageSample* or OpImageSparseSample* instructions that includes a LOD bias or any offset values, in any shader stage

If the robust buffer access feature is not enabled, and if the VkPipeline object bound to the pipeline bind point used by this command accesses a uniform buffer, it must not access values outside of the range of the buffer as specified in the descriptor set bound to the same pipeline bind point

If the robust buffer access feature is not enabled, and if the VkPipeline object bound to the pipeline bind point used by this command accesses a storage buffer, it must not access values outside of the range of the buffer as specified in the descriptor set bound to the same pipeline bind point

The current render pass must be compatible with the renderPass member of the VkGraphicsPipelineCreateInfo structure specified when creating the VkPipeline bound to VK_PIPELINE_BIND_POINT_GRAPHICS.

The subpass index of the current render pass must be equal to the subpass member of the VkGraphicsPipelineCreateInfo structure specified when creating the VkPipeline bound to VK_PIPELINE_BIND_POINT_GRAPHICS.

Every input attachment used by the current subpass must be bound to the pipeline via a descriptor set

Image subresources used as attachments in the current render pass must not be accessed in any way other than as an attachment by this command.

If buffer is non-sparse then it must be bound completely and contiguously to a single VkDeviceMemory object

buffer must have been created with the VK_BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set

offset must be a multiple of 4

If countBuffer is non-sparse then it must be bound completely and contiguously to a single VkDeviceMemory object

countBuffer must have been created with the VK_BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set

countBufferOffset must be a multiple of 4

The count stored in countBuffer must be less than or equal to VkPhysicalDeviceLimits::maxDrawIndirectCount

stride must be a multiple of 4 and must be greater than or equal to sizeof(VkDrawMeshTasksIndirectCommandNV)

If maxDrawCount is greater than or equal to 1, (stride × (maxDrawCount - 1) + offset + sizeof(VkDrawMeshTasksIndirectCommandNV)) must be less than or equal to the size of buffer

If the count stored in countBuffer is equal to 1, (offset + sizeof(VkDrawMeshTasksIndirectCommandNV)) must be less than or equal to the size of buffer

If the count stored in countBuffer is greater than 1, (stride × (drawCount - 1) + offset + sizeof(VkDrawMeshTasksIndirectCommandNV)) must be less than or equal to the size of buffer

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

buffer must be a valid VkBuffer handle

countBuffer must be a valid VkBuffer handle

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics operations

This command must only be called inside of a render pass instance

Each of buffer, commandBuffer, and countBuffer must have been created, allocated, or retrieved from the same VkDevice

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Inside          | Graphics        | Graphics        |
| Secondary       |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkBuffer, VkCommandBuffer, VkDeviceSize

cmdDrawMeshTasksIndirectNV :: CommandBuffer -> Buffer -> DeviceSize -> Word32 -> Word32 -> IO ()

vkCmdDrawMeshTasksIndirectNV - Issue an indirect mesh tasks draw into a command buffer

Parameters

commandBuffer is the command buffer into which the command is recorded.

buffer is the buffer containing draw parameters.

offset is the byte offset into buffer where parameters begin.

drawCount is the number of draws to execute, and can be zero.

stride is the byte stride between successive sets of draw parameters.

Description

vkCmdDrawMeshTasksIndirectNV behaves similarly to vkCmdDrawMeshTasksNV except that the parameters are read by the device from a buffer during execution. drawCount draws are executed by the command, with parameters taken from buffer starting at offset and increasing by stride bytes for each successive draw. The parameters of each draw are encoded in an array of VkDrawMeshTasksIndirectCommandNV structures. If drawCount is less than or equal to one, stride is ignored.

Valid Usage

If a VkImageView is sampled with VK_FILTER_LINEAR as a result of this command, then the image view’s format features must contain VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_LINEAR_BIT

If a VkImageView is accessed using atomic operations as a result of this command, then the image view’s format features must contain VK_FORMAT_FEATURE_STORAGE_IMAGE_ATOMIC_BIT

For each set n that is statically used by the VkPipeline bound to the pipeline bind point used by this command, a descriptor set must have been bound to n at the same pipeline bind point, with a VkPipelineLayout that is compatible for set n, with the VkPipelineLayout used to create the current VkPipeline, as described in ???

For each push constant that is statically used by the VkPipeline bound to the pipeline bind point used by this command, a push constant value must have been set for the same pipeline bind point, with a VkPipelineLayout that is compatible for push constants, with the VkPipelineLayout used to create the current VkPipeline, as described in ???

Descriptors in each bound descriptor set, specified via vkCmdBindDescriptorSets, must be valid if they are statically used by the VkPipeline bound to the pipeline bind point used by this command

A valid pipeline must be bound to the pipeline bind point used by this command

If the VkPipeline object bound to the pipeline bind point used by this command requires any dynamic state, that state must have been set for commandBuffer

If the VkPipeline object bound to the pipeline bind point used by this command accesses a VkSampler object that uses unnormalized coordinates, that sampler must not be used to sample from any VkImage with a VkImageView of the type VK_IMAGE_VIEW_TYPE_3D, VK_IMAGE_VIEW_TYPE_CUBE, VK_IMAGE_VIEW_TYPE_1D_ARRAY, VK_IMAGE_VIEW_TYPE_2D_ARRAY or VK_IMAGE_VIEW_TYPE_CUBE_ARRAY, in any shader stage

If the VkPipeline object bound to the pipeline bind point used by this command accesses a VkSampler object that uses unnormalized coordinates, that sampler must not be used with any of the SPIR-V OpImageSample* or OpImageSparseSample* instructions with ImplicitLod, Dref or Proj in their name, in any shader stage

If the VkPipeline object bound to the pipeline bind point used by this command accesses a VkSampler object that uses unnormalized coordinates, that sampler must not be used with any of the SPIR-V OpImageSample* or OpImageSparseSample* instructions that includes a LOD bias or any offset values, in any shader stage

If the robust buffer access feature is not enabled, and if the VkPipeline object bound to the pipeline bind point used by this command accesses a uniform buffer, it must not access values outside of the range of the buffer as specified in the descriptor set bound to the same pipeline bind point

If the robust buffer access feature is not enabled, and if the VkPipeline object bound to the pipeline bind point used by this command accesses a storage buffer, it must not access values outside of the range of the buffer as specified in the descriptor set bound to the same pipeline bind point

The current render pass must be compatible with the renderPass member of the VkGraphicsPipelineCreateInfo structure specified when creating the VkPipeline bound to VK_PIPELINE_BIND_POINT_GRAPHICS.

The subpass index of the current render pass must be equal to the subpass member of the VkGraphicsPipelineCreateInfo structure specified when creating the VkPipeline bound to VK_PIPELINE_BIND_POINT_GRAPHICS.

Every input attachment used by the current subpass must be bound to the pipeline via a descriptor set

Image subresources used as attachments in the current render pass must not be accessed in any way other than as an attachment by this command.

If buffer is non-sparse then it must be bound completely and contiguously to a single VkDeviceMemory object

buffer must have been created with the VK_BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set

offset must be a multiple of 4

If the multi-draw indirect feature is not enabled, drawCount must be 0 or 1

drawCount must be less than or equal to VkPhysicalDeviceLimits::maxDrawIndirectCount

If drawCount is greater than 1, stride must be a multiple of 4 and must be greater than or equal to sizeof(VkDrawMeshTasksIndirectCommandNV)

If drawCount is equal to 1, (offset + sizeof(VkDrawMeshTasksIndirectCommandNV)) must be less than or equal to the size of buffer

If drawCount is greater than 1, (stride × (drawCount - 1) + offset + sizeof(VkDrawMeshTasksIndirectCommandNV)) must be less than or equal to the size of buffer

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

buffer must be a valid VkBuffer handle

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics operations

This command must only be called inside of a render pass instance

Both of buffer, and commandBuffer must have been created, allocated, or retrieved from the same VkDevice

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Inside          | Graphics        | Graphics        |
| Secondary       |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkBuffer, VkCommandBuffer, VkDeviceSize

cmdDrawMeshTasksNV :: CommandBuffer -> Word32 -> Word32 -> IO ()

vkCmdDrawMeshTasksNV - Draw mesh task work items

Parameters

commandBuffer is the command buffer into which the command will be recorded.

taskCount is the number of local workgroups to dispatch in the X dimension. Y and Z dimension are implicitly set to one.

firstTask is the X component of the first workgroup ID.

Description

When the command is executed, a global workgroup consisting of taskCount local workgroups is assembled.

Valid Usage

If a VkImageView is sampled with VK_FILTER_LINEAR as a result of this command, then the image view’s format features must contain VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_LINEAR_BIT

If a VkImageView is accessed using atomic operations as a result of this command, then the image view’s format features must contain VK_FORMAT_FEATURE_STORAGE_IMAGE_ATOMIC_BIT

For each set n that is statically used by the VkPipeline bound to the pipeline bind point used by this command, a descriptor set must have been bound to n at the same pipeline bind point, with a VkPipelineLayout that is compatible for set n, with the VkPipelineLayout used to create the current VkPipeline, as described in ???

For each push constant that is statically used by the VkPipeline bound to the pipeline bind point used by this command, a push constant value must have been set for the same pipeline bind point, with a VkPipelineLayout that is compatible for push constants, with the VkPipelineLayout used to create the current VkPipeline, as described in ???

Descriptors in each bound descriptor set, specified via vkCmdBindDescriptorSets, must be valid if they are statically used by the VkPipeline bound to the pipeline bind point used by this command

A valid pipeline must be bound to the pipeline bind point used by this command

If the VkPipeline object bound to the pipeline bind point used by this command requires any dynamic state, that state must have been set for commandBuffer

If the VkPipeline object bound to the pipeline bind point used by this command accesses a VkSampler object that uses unnormalized coordinates, that sampler must not be used to sample from any VkImage with a VkImageView of the type VK_IMAGE_VIEW_TYPE_3D, VK_IMAGE_VIEW_TYPE_CUBE, VK_IMAGE_VIEW_TYPE_1D_ARRAY, VK_IMAGE_VIEW_TYPE_2D_ARRAY or VK_IMAGE_VIEW_TYPE_CUBE_ARRAY, in any shader stage

If the VkPipeline object bound to the pipeline bind point used by this command accesses a VkSampler object that uses unnormalized coordinates, that sampler must not be used with any of the SPIR-V OpImageSample* or OpImageSparseSample* instructions with ImplicitLod, Dref or Proj in their name, in any shader stage

If the VkPipeline object bound to the pipeline bind point used by this command accesses a VkSampler object that uses unnormalized coordinates, that sampler must not be used with any of the SPIR-V OpImageSample* or OpImageSparseSample* instructions that includes a LOD bias or any offset values, in any shader stage

If the robust buffer access feature is not enabled, and if the VkPipeline object bound to the pipeline bind point used by this command accesses a uniform buffer, it must not access values outside of the range of the buffer as specified in the descriptor set bound to the same pipeline bind point

If the robust buffer access feature is not enabled, and if the VkPipeline object bound to the pipeline bind point used by this command accesses a storage buffer, it must not access values outside of the range of the buffer as specified in the descriptor set bound to the same pipeline bind point

The current render pass must be compatible with the renderPass member of the VkGraphicsPipelineCreateInfo structure specified when creating the VkPipeline bound to VK_PIPELINE_BIND_POINT_GRAPHICS.

The subpass index of the current render pass must be equal to the subpass member of the VkGraphicsPipelineCreateInfo structure specified when creating the VkPipeline bound to VK_PIPELINE_BIND_POINT_GRAPHICS.

Every input attachment used by the current subpass must be bound to the pipeline via a descriptor set

Image subresources used as attachments in the current render pass must not be accessed in any way other than as an attachment by this command.

taskCount must be less than or equal to VkPhysicalDeviceMeshShaderPropertiesNV::maxDrawMeshTasksCount

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics operations

This command must only be called inside of a render pass instance

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Inside          | Graphics        | Graphics        |
| Secondary       |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer

pattern NV_MESH_SHADER_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NV_MESH_SHADER_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_MESH_SHADER_FEATURES_NV :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_MESH_SHADER_PROPERTIES_NV :: a ~ StructureType => a

pattern SHADER_STAGE_TASK_BIT_NV :: a ~ ShaderStageFlagBits => a

pattern SHADER_STAGE_MESH_BIT_NV :: a ~ ShaderStageFlagBits => a

pattern PIPELINE_STAGE_TASK_SHADER_BIT_NV :: a ~ PipelineStageFlagBits => a

pattern PIPELINE_STAGE_MESH_SHADER_BIT_NV :: a ~ PipelineStageFlagBits => a