Graphics.Vulkan.Extensions.VK_NV_representative_fragment_test

Documentation

withCStructPhysicalDeviceRepresentativeFragmentTestFeaturesNV :: PhysicalDeviceRepresentativeFragmentTestFeaturesNV -> (VkPhysicalDeviceRepresentativeFragmentTestFeaturesNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceRepresentativeFragmentTestFeaturesNV and marshal a PhysicalDeviceRepresentativeFragmentTestFeaturesNV into it. The VkPhysicalDeviceRepresentativeFragmentTestFeaturesNV is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceRepresentativeFragmentTestFeaturesNV :: VkPhysicalDeviceRepresentativeFragmentTestFeaturesNV -> IO PhysicalDeviceRepresentativeFragmentTestFeaturesNV

A function to read a VkPhysicalDeviceRepresentativeFragmentTestFeaturesNV and all additional structures in the pointer chain into a PhysicalDeviceRepresentativeFragmentTestFeaturesNV.

data PhysicalDeviceRepresentativeFragmentTestFeaturesNV

VkPhysicalDeviceRepresentativeFragmentTestFeaturesNV - Structure describing the representative fragment test features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceRepresentativeFragmentTestFeaturesNV structure describe the following features:

Description

If the VkPhysicalDeviceRepresentativeFragmentTestFeaturesNV structure is included in the pNext chain of VkPhysicalDeviceFeatures2KHR, it is filled with values indicating whether the feature is supported. VkPhysicalDeviceRepresentativeFragmentTestFeaturesNV can also be used in the pNext chain of VkDeviceCreateInfo to enable the feature.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceRepresentativeFragmentTestFeaturesNV

Fields

next :: Maybe SomeVkStruct

representativeFragmentTest :: Bool

withCStructPipelineRepresentativeFragmentTestStateCreateInfoNV :: PipelineRepresentativeFragmentTestStateCreateInfoNV -> (VkPipelineRepresentativeFragmentTestStateCreateInfoNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkPipelineRepresentativeFragmentTestStateCreateInfoNV and marshal a PipelineRepresentativeFragmentTestStateCreateInfoNV into it. The VkPipelineRepresentativeFragmentTestStateCreateInfoNV is only valid inside the provided computation and must not be returned out of it.

fromCStructPipelineRepresentativeFragmentTestStateCreateInfoNV :: VkPipelineRepresentativeFragmentTestStateCreateInfoNV -> IO PipelineRepresentativeFragmentTestStateCreateInfoNV

A function to read a VkPipelineRepresentativeFragmentTestStateCreateInfoNV and all additional structures in the pointer chain into a PipelineRepresentativeFragmentTestStateCreateInfoNV.

data PipelineRepresentativeFragmentTestStateCreateInfoNV

VkPipelineRepresentativeFragmentTestStateCreateInfoNV - Structure specifying representative fragment test

Description

If this structure is not present, representativeFragmentTestEnable is considered to be VK_FALSE, and the representative fragment test is disabled.

If early fragment tests are not enabled in the active fragment shader, the representative fragment shader test has no effect, even if enabled.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PipelineRepresentativeFragmentTestStateCreateInfoNV

Fields

next :: Maybe SomeVkStruct

representativeFragmentTestEnable :: Bool

pattern NV_REPRESENTATIVE_FRAGMENT_TEST_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NV_REPRESENTATIVE_FRAGMENT_TEST_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_REPRESENTATIVE_FRAGMENT_TEST_FEATURES_NV :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PIPELINE_REPRESENTATIVE_FRAGMENT_TEST_STATE_CREATE_INFO_NV :: a ~ StructureType => a