Graphics.Vulkan.Extensions.VK_NV_shader_image_footprint

Documentation

withCStructPhysicalDeviceShaderImageFootprintFeaturesNV :: PhysicalDeviceShaderImageFootprintFeaturesNV -> (VkPhysicalDeviceShaderImageFootprintFeaturesNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceShaderImageFootprintFeaturesNV and marshal a PhysicalDeviceShaderImageFootprintFeaturesNV into it. The VkPhysicalDeviceShaderImageFootprintFeaturesNV is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceShaderImageFootprintFeaturesNV :: VkPhysicalDeviceShaderImageFootprintFeaturesNV -> IO PhysicalDeviceShaderImageFootprintFeaturesNV

A function to read a VkPhysicalDeviceShaderImageFootprintFeaturesNV and all additional structures in the pointer chain into a PhysicalDeviceShaderImageFootprintFeaturesNV.

data PhysicalDeviceShaderImageFootprintFeaturesNV

VkPhysicalDeviceShaderImageFootprintFeaturesNV - Structure describing shader image footprint features that can be supported by an implementation

Description

See Texel Footprint Evaluation for more information.

If the VkPhysicalDeviceShaderImageFootprintFeaturesNV structure is included in the pNext chain of VkPhysicalDeviceFeatures2KHR, it is filled with values indicating whether each feature is supported. VkPhysicalDeviceShaderImageFootprintFeaturesNV can also be used in the pNext chain of VkDeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceShaderImageFootprintFeaturesNV

Fields

next :: Maybe SomeVkStruct

imageFootprint :: Bool

pattern NV_SHADER_IMAGE_FOOTPRINT_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NV_SHADER_IMAGE_FOOTPRINT_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_IMAGE_FOOTPRINT_FEATURES_NV :: a ~ StructureType => a