Graphics.Vulkan.Extensions.VK_NV_shading_rate_image

Documentation

withCStructCoarseSampleLocationNV :: CoarseSampleLocationNV -> (VkCoarseSampleLocationNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkCoarseSampleLocationNV and marshal a CoarseSampleLocationNV into it. The VkCoarseSampleLocationNV is only valid inside the provided computation and must not be returned out of it.

fromCStructCoarseSampleLocationNV :: VkCoarseSampleLocationNV -> IO CoarseSampleLocationNV

A function to read a VkCoarseSampleLocationNV and all additional structures in the pointer chain into a CoarseSampleLocationNV.

data CoarseSampleLocationNV

VkCoarseSampleLocationNV - Structure specifying parameters controlling shading rate image usage

Valid Usage

See Also

VkCoarseSampleOrderCustomNV

Constructors

CoarseSampleLocationNV

Fields

pixelX :: Word32

pixelY :: Word32

sample :: Word32

withCStructCoarseSampleOrderCustomNV :: CoarseSampleOrderCustomNV -> (VkCoarseSampleOrderCustomNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkCoarseSampleOrderCustomNV and marshal a CoarseSampleOrderCustomNV into it. The VkCoarseSampleOrderCustomNV is only valid inside the provided computation and must not be returned out of it.

fromCStructCoarseSampleOrderCustomNV :: VkCoarseSampleOrderCustomNV -> IO CoarseSampleOrderCustomNV

A function to read a VkCoarseSampleOrderCustomNV and all additional structures in the pointer chain into a CoarseSampleOrderCustomNV.

data CoarseSampleOrderCustomNV

VkCoarseSampleOrderCustomNV - Structure specifying parameters controlling shading rate image usage

Description

When using a custom sample ordering, element i in pSampleLocations specifies a specific pixel and per-pixel coverage sample number that corresponds to the coverage sample numbered i in the multi-pixel fragment.

Valid Usage

shadingRate must be a shading rate that generates fragments with more than one pixel.

sampleCount must correspond to a sample count enumerated in VkSampleCountFlags whose corresponding bit is set in VkPhysicalDeviceLimits::framebufferNoAttachmentsSampleCounts.

sampleLocationCount must be equal to the product of sampleCount, the fragment width for shadingRate, and the fragment height for shadingRate.

sampleLocationCount must be less than or equal to the value of VkPhysicalDeviceShadingRateImagePropertiesNV::shadingRateMaxCoarseSamples.

The array pSampleLocations must contain exactly one entry for every combination of valid values for pixelX, pixelY, and sample in the structure VkCoarseSampleOrderCustomNV.

Valid Usage (Implicit)

shadingRate must be a valid VkShadingRatePaletteEntryNV value

pSampleLocations must be a valid pointer to an array of sampleLocationCount VkCoarseSampleLocationNV structures

sampleLocationCount must be greater than 0

See Also

VkCoarseSampleLocationNV, VkPipelineViewportCoarseSampleOrderStateCreateInfoNV, VkShadingRatePaletteEntryNV, vkCmdSetCoarseSampleOrderNV

Constructors

CoarseSampleOrderCustomNV

Fields

shadingRate :: ShadingRatePaletteEntryNV

sampleCount :: Word32

sampleLocations :: Vector CoarseSampleLocationNV

type CoarseSampleOrderTypeNV = VkCoarseSampleOrderTypeNV

VkCoarseSampleOrderTypeNV - Shading rate image sample ordering types

See Also

VkPipelineViewportCoarseSampleOrderStateCreateInfoNV, vkCmdSetCoarseSampleOrderNV

pattern COARSE_SAMPLE_ORDER_TYPE_DEFAULT_NV :: a ~ CoarseSampleOrderTypeNV => a

VK_COARSE_SAMPLE_ORDER_TYPE_DEFAULT_NV specifies that coverage samples will be ordered in an implementation-dependent manner.

pattern COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV :: a ~ CoarseSampleOrderTypeNV => a

VK_COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV specifies that coverage samples will be ordered according to the array of custom orderings provided in either the pCustomSampleOrders member of VkPipelineViewportCoarseSampleOrderStateCreateInfoNV or the pCustomSampleOrders member of vkCmdSetCoarseSampleOrderNV.

pattern COARSE_SAMPLE_ORDER_TYPE_PIXEL_MAJOR_NV :: a ~ CoarseSampleOrderTypeNV => a

VK_COARSE_SAMPLE_ORDER_TYPE_PIXEL_MAJOR_NV specifies that coverage samples will be ordered sequentially, sorted first by pixel coordinate (in row-major order) and then by coverage sample number.

pattern COARSE_SAMPLE_ORDER_TYPE_SAMPLE_MAJOR_NV :: a ~ CoarseSampleOrderTypeNV => a

VK_COARSE_SAMPLE_ORDER_TYPE_SAMPLE_MAJOR_NV specifies that coverage samples will be ordered sequentially, sorted first by coverage sample number and then by pixel coordinate (in row-major order).

withCStructPhysicalDeviceShadingRateImageFeaturesNV :: PhysicalDeviceShadingRateImageFeaturesNV -> (VkPhysicalDeviceShadingRateImageFeaturesNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceShadingRateImageFeaturesNV and marshal a PhysicalDeviceShadingRateImageFeaturesNV into it. The VkPhysicalDeviceShadingRateImageFeaturesNV is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceShadingRateImageFeaturesNV :: VkPhysicalDeviceShadingRateImageFeaturesNV -> IO PhysicalDeviceShadingRateImageFeaturesNV

A function to read a VkPhysicalDeviceShadingRateImageFeaturesNV and all additional structures in the pointer chain into a PhysicalDeviceShadingRateImageFeaturesNV.

data PhysicalDeviceShadingRateImageFeaturesNV

VkPhysicalDeviceShadingRateImageFeaturesNV - Structure describing shading rate image features that can be supported by an implementation

Members

The members of the VkPhysicalDeviceShadingRateImageFeaturesNV structure describe the following features:

Description

See Shading Rate Image for more information.

If the VkPhysicalDeviceShadingRateImageFeaturesNV structure is included in the pNext chain of VkPhysicalDeviceFeatures2KHR, it is filled with values indicating whether the feature is supported. VkPhysicalDeviceShadingRateImageFeaturesNV can also be used in the pNext chain of VkDeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType

Constructors

PhysicalDeviceShadingRateImageFeaturesNV

Fields

next :: Maybe SomeVkStruct

shadingRateImage :: Bool

shadingRateCoarseSampleOrder :: Bool

withCStructPhysicalDeviceShadingRateImagePropertiesNV :: PhysicalDeviceShadingRateImagePropertiesNV -> (VkPhysicalDeviceShadingRateImagePropertiesNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkPhysicalDeviceShadingRateImagePropertiesNV and marshal a PhysicalDeviceShadingRateImagePropertiesNV into it. The VkPhysicalDeviceShadingRateImagePropertiesNV is only valid inside the provided computation and must not be returned out of it.

fromCStructPhysicalDeviceShadingRateImagePropertiesNV :: VkPhysicalDeviceShadingRateImagePropertiesNV -> IO PhysicalDeviceShadingRateImagePropertiesNV

A function to read a VkPhysicalDeviceShadingRateImagePropertiesNV and all additional structures in the pointer chain into a PhysicalDeviceShadingRateImagePropertiesNV.

data PhysicalDeviceShadingRateImagePropertiesNV

VkPhysicalDeviceShadingRateImagePropertiesNV - Structure describing shading rate image limits that can be supported by an implementation

Members

The members of the VkPhysicalDeviceShadingRateImagePropertiesNV structure describe the following implementation-dependent properties related to the shading rate image feature:

Description

If the VkPhysicalDeviceShadingRateImagePropertiesNV structure is included in the pNext chain of VkPhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

VkExtent2D, VkStructureType

Constructors

PhysicalDeviceShadingRateImagePropertiesNV

Fields

next :: Maybe SomeVkStruct

shadingRateTexelSize :: Extent2D

shadingRatePaletteSize :: Word32

shadingRateMaxCoarseSamples :: Word32

withCStructPipelineViewportCoarseSampleOrderStateCreateInfoNV :: PipelineViewportCoarseSampleOrderStateCreateInfoNV -> (VkPipelineViewportCoarseSampleOrderStateCreateInfoNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkPipelineViewportCoarseSampleOrderStateCreateInfoNV and marshal a PipelineViewportCoarseSampleOrderStateCreateInfoNV into it. The VkPipelineViewportCoarseSampleOrderStateCreateInfoNV is only valid inside the provided computation and must not be returned out of it.

fromCStructPipelineViewportCoarseSampleOrderStateCreateInfoNV :: VkPipelineViewportCoarseSampleOrderStateCreateInfoNV -> IO PipelineViewportCoarseSampleOrderStateCreateInfoNV

A function to read a VkPipelineViewportCoarseSampleOrderStateCreateInfoNV and all additional structures in the pointer chain into a PipelineViewportCoarseSampleOrderStateCreateInfoNV.

data PipelineViewportCoarseSampleOrderStateCreateInfoNV

VkPipelineViewportCoarseSampleOrderStateCreateInfoNV - Structure specifying parameters controlling sample order in coarse fragments

Description

If this structure is not present, sampleOrderType is considered to be VK_COARSE_SAMPLE_ORDER_TYPE_DEFAULT_NV.

If sampleOrderType is VK_COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV, the coverage sample order used for any combination of fragment area and coverage sample count not enumerated in pCustomSampleOrders will be identical to that used for VK_COARSE_SAMPLE_ORDER_TYPE_DEFAULT_NV.

If the pipeline was created with VK_DYNAMIC_STATE_VIEWPORT_COARSE_SAMPLE_ORDER_NV, the contents of this structure (if present) are ignored, and the coverage sample order is instead specified by vkCmdSetCoarseSampleOrderNV.

Valid Usage

If sampleOrderType is not VK_COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV, customSamplerOrderCount must be 0

The array pCustomSampleOrders must not contain two structures with matching values for both the shadingRate and sampleCount members.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_COARSE_SAMPLE_ORDER_STATE_CREATE_INFO_NV

sampleOrderType must be a valid VkCoarseSampleOrderTypeNV value

If customSampleOrderCount is not 0, pCustomSampleOrders must be a valid pointer to an array of customSampleOrderCount valid VkCoarseSampleOrderCustomNV structures

See Also

VkCoarseSampleOrderCustomNV, VkCoarseSampleOrderTypeNV, VkStructureType

Constructors

PipelineViewportCoarseSampleOrderStateCreateInfoNV

Fields

next :: Maybe SomeVkStruct

sampleOrderType :: CoarseSampleOrderTypeNV

customSampleOrders :: Vector CoarseSampleOrderCustomNV

withCStructPipelineViewportShadingRateImageStateCreateInfoNV :: PipelineViewportShadingRateImageStateCreateInfoNV -> (VkPipelineViewportShadingRateImageStateCreateInfoNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkPipelineViewportShadingRateImageStateCreateInfoNV and marshal a PipelineViewportShadingRateImageStateCreateInfoNV into it. The VkPipelineViewportShadingRateImageStateCreateInfoNV is only valid inside the provided computation and must not be returned out of it.

fromCStructPipelineViewportShadingRateImageStateCreateInfoNV :: VkPipelineViewportShadingRateImageStateCreateInfoNV -> IO PipelineViewportShadingRateImageStateCreateInfoNV

A function to read a VkPipelineViewportShadingRateImageStateCreateInfoNV and all additional structures in the pointer chain into a PipelineViewportShadingRateImageStateCreateInfoNV.

data PipelineViewportShadingRateImageStateCreateInfoNV

VkPipelineViewportShadingRateImageStateCreateInfoNV - Structure specifying parameters controlling shading rate image usage

Description

If this structure is not present, shadingRateImageEnable is considered to be VK_FALSE, and the shading rate image and palettes are not used.

Valid Usage

If the multiple viewports feature is not enabled, viewportCount must be 0 or 1

viewportCount must be less than or equal to VkPhysicalDeviceLimits::maxViewports

If shadingRateImageEnable is VK_TRUE, viewportCount must be equal to the viewportCount member of VkPipelineViewportStateCreateInfo

If no element of the pDynamicStates member of pDynamicState is VK_DYNAMIC_STATE_VIEWPORT_SHADING_RATE_PALETTE_NV, pShadingRatePalettes must be a valid pointer to an array of viewportCount VkShadingRatePaletteNV structures

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_SHADING_RATE_IMAGE_STATE_CREATE_INFO_NV

If viewportCount is not 0, and pShadingRatePalettes is not NULL, pShadingRatePalettes must be a valid pointer to an array of viewportCount valid VkShadingRatePaletteNV structures

See Also

VkBool32, VkShadingRatePaletteNV, VkStructureType

Constructors

PipelineViewportShadingRateImageStateCreateInfoNV

Fields

next :: Maybe SomeVkStruct

shadingRateImageEnable :: Bool

shadingRatePalettes :: Maybe (Vector ShadingRatePaletteNV)

type ShadingRatePaletteEntryNV = VkShadingRatePaletteEntryNV

VkShadingRatePaletteEntryNV - Shading rate image palette entry types

Description

The following table indicates the width and height (in pixels) of each fragment generated using the indicated shading rate, as well as the maximum number of fragment shader invocations launched for each fragment. When processing regions of a primitive that have a shading rate of VK_SHADING_RATE_PALETTE_ENTRY_NO_INVOCATIONS_NV, no fragments will be generated in that region.

+-----------------+-----------------+-----------------+-----------------+
| Shading Rate    | Width           | Height          | Invocations     |
+=================+=================+=================+=================+
| 'Graphics.Vulka | 0               | 0               | 0               |
| n.C.Extensions. |                 |                 |                 |
| VK_NV_shading_r |                 |                 |                 |
| ate_image.VK_SH |                 |                 |                 |
| ADING_RATE_PALE |                 |                 |                 |
| TTE_ENTRY_NO_IN |                 |                 |                 |
| VOCATIONS_NV'   |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+
| 'Graphics.Vulka | 1               | 1               | 16              |
| n.C.Extensions. |                 |                 |                 |
| VK_NV_shading_r |                 |                 |                 |
| ate_image.VK_SH |                 |                 |                 |
| ADING_RATE_PALE |                 |                 |                 |
| TTE_ENTRY_16_IN |                 |                 |                 |
| VOCATIONS_PER_P |                 |                 |                 |
| IXEL_NV'        |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+
| 'Graphics.Vulka | 1               | 1               | 8               |
| n.C.Extensions. |                 |                 |                 |
| VK_NV_shading_r |                 |                 |                 |
| ate_image.VK_SH |                 |                 |                 |
| ADING_RATE_PALE |                 |                 |                 |
| TTE_ENTRY_8_INV |                 |                 |                 |
| OCATIONS_PER_PI |                 |                 |                 |
| XEL_NV'         |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+
| 'Graphics.Vulka | 1               | 1               | 4               |
| n.C.Extensions. |                 |                 |                 |
| VK_NV_shading_r |                 |                 |                 |
| ate_image.VK_SH |                 |                 |                 |
| ADING_RATE_PALE |                 |                 |                 |
| TTE_ENTRY_4_INV |                 |                 |                 |
| OCATIONS_PER_PI |                 |                 |                 |
| XEL_NV'         |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+
| 'Graphics.Vulka | 1               | 1               | 2               |
| n.C.Extensions. |                 |                 |                 |
| VK_NV_shading_r |                 |                 |                 |
| ate_image.VK_SH |                 |                 |                 |
| ADING_RATE_PALE |                 |                 |                 |
| TTE_ENTRY_2_INV |                 |                 |                 |
| OCATIONS_PER_PI |                 |                 |                 |
| XEL_NV'         |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+
| 'Graphics.Vulka | 1               | 1               | 1               |
| n.C.Extensions. |                 |                 |                 |
| VK_NV_shading_r |                 |                 |                 |
| ate_image.VK_SH |                 |                 |                 |
| ADING_RATE_PALE |                 |                 |                 |
| TTE_ENTRY_1_INV |                 |                 |                 |
| OCATION_PER_PIX |                 |                 |                 |
| EL_NV'          |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+
| 'Graphics.Vulka | 2               | 1               | 1               |
| n.C.Extensions. |                 |                 |                 |
| VK_NV_shading_r |                 |                 |                 |
| ate_image.VK_SH |                 |                 |                 |
| ADING_RATE_PALE |                 |                 |                 |
| TTE_ENTRY_1_INV |                 |                 |                 |
| OCATION_PER_2X1 |                 |                 |                 |
| _PIXELS_NV'     |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+
| 'Graphics.Vulka | 1               | 2               | 1               |
| n.C.Extensions. |                 |                 |                 |
| VK_NV_shading_r |                 |                 |                 |
| ate_image.VK_SH |                 |                 |                 |
| ADING_RATE_PALE |                 |                 |                 |
| TTE_ENTRY_1_INV |                 |                 |                 |
| OCATION_PER_1X2 |                 |                 |                 |
| _PIXELS_NV'     |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+
| 'Graphics.Vulka | 2               | 2               | 1               |
| n.C.Extensions. |                 |                 |                 |
| VK_NV_shading_r |                 |                 |                 |
| ate_image.VK_SH |                 |                 |                 |
| ADING_RATE_PALE |                 |                 |                 |
| TTE_ENTRY_1_INV |                 |                 |                 |
| OCATION_PER_2X2 |                 |                 |                 |
| _PIXELS_NV'     |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+
| 'Graphics.Vulka | 4               | 2               | 1               |
| n.C.Extensions. |                 |                 |                 |
| VK_NV_shading_r |                 |                 |                 |
| ate_image.VK_SH |                 |                 |                 |
| ADING_RATE_PALE |                 |                 |                 |
| TTE_ENTRY_1_INV |                 |                 |                 |
| OCATION_PER_4X2 |                 |                 |                 |
| _PIXELS_NV'     |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+
| 'Graphics.Vulka | 2               | 4               | 1               |
| n.C.Extensions. |                 |                 |                 |
| VK_NV_shading_r |                 |                 |                 |
| ate_image.VK_SH |                 |                 |                 |
| ADING_RATE_PALE |                 |                 |                 |
| TTE_ENTRY_1_INV |                 |                 |                 |
| OCATION_PER_2X4 |                 |                 |                 |
| _PIXELS_NV'     |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+
| 'Graphics.Vulka | 4               | 4               | 1               |
| n.C.Extensions. |                 |                 |                 |
| VK_NV_shading_r |                 |                 |                 |
| ate_image.VK_SH |                 |                 |                 |
| ADING_RATE_PALE |                 |                 |                 |
| TTE_ENTRY_1_INV |                 |                 |                 |
| OCATION_PER_4X4 |                 |                 |                 |
| _PIXELS_NV'     |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCoarseSampleOrderCustomNV, VkShadingRatePaletteNV

pattern SHADING_RATE_PALETTE_ENTRY_NO_INVOCATIONS_NV :: a ~ ShadingRatePaletteEntryNV => a

pattern SHADING_RATE_PALETTE_ENTRY_16_INVOCATIONS_PER_PIXEL_NV :: a ~ ShadingRatePaletteEntryNV => a

pattern SHADING_RATE_PALETTE_ENTRY_8_INVOCATIONS_PER_PIXEL_NV :: a ~ ShadingRatePaletteEntryNV => a

pattern SHADING_RATE_PALETTE_ENTRY_4_INVOCATIONS_PER_PIXEL_NV :: a ~ ShadingRatePaletteEntryNV => a

pattern SHADING_RATE_PALETTE_ENTRY_2_INVOCATIONS_PER_PIXEL_NV :: a ~ ShadingRatePaletteEntryNV => a

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_PIXEL_NV :: a ~ ShadingRatePaletteEntryNV => a

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_2X1_PIXELS_NV :: a ~ ShadingRatePaletteEntryNV => a

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_1X2_PIXELS_NV :: a ~ ShadingRatePaletteEntryNV => a

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_2X2_PIXELS_NV :: a ~ ShadingRatePaletteEntryNV => a

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_4X2_PIXELS_NV :: a ~ ShadingRatePaletteEntryNV => a

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_2X4_PIXELS_NV :: a ~ ShadingRatePaletteEntryNV => a

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_4X4_PIXELS_NV :: a ~ ShadingRatePaletteEntryNV => a

withCStructShadingRatePaletteNV :: ShadingRatePaletteNV -> (VkShadingRatePaletteNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkShadingRatePaletteNV and marshal a ShadingRatePaletteNV into it. The VkShadingRatePaletteNV is only valid inside the provided computation and must not be returned out of it.

fromCStructShadingRatePaletteNV :: VkShadingRatePaletteNV -> IO ShadingRatePaletteNV

A function to read a VkShadingRatePaletteNV and all additional structures in the pointer chain into a ShadingRatePaletteNV.

data ShadingRatePaletteNV

VkShadingRatePaletteNV - Structure specifying a single shading rate palette

Valid Usage (Implicit)

See Also

VkPipelineViewportShadingRateImageStateCreateInfoNV, VkShadingRatePaletteEntryNV, vkCmdSetViewportShadingRatePaletteNV

Constructors

ShadingRatePaletteNV

Fields

shadingRatePaletteEntries :: Vector ShadingRatePaletteEntryNV

cmdBindShadingRateImageNV :: CommandBuffer -> ImageView -> ImageLayout -> IO ()

vkCmdBindShadingRateImageNV - Bind a shading rate image on a command buffer

Parameters

commandBuffer is the command buffer into which the command will be recorded.

imageView is an image view handle that specifies the shading rate image. imageView may be set to VK_NULL_HANDLE, which is equivalent to specifying a view of an image filled with zero values.

imageLayout is the layout that the image subresources accessible from imageView will be in when the shading rate image is accessed.

Valid Usage

The shading rate image feature must be enabled.

If imageView is not VK_NULL_HANDLE, it must be a valid VkImageView handle of type VK_IMAGE_VIEW_TYPE_2D or VK_IMAGE_VIEW_TYPE_2D_ARRAY.

If imageView is not VK_NULL_HANDLE, it must have a format of VK_FORMAT_R8_UINT.

If imageView is not VK_NULL_HANDLE, it must have been created with a usage value including VK_IMAGE_USAGE_SHADING_RATE_IMAGE_BIT_NV

If imageView is not VK_NULL_HANDLE, imageLayout must match the actual VkImageLayout of each subresource accessible from imageView at the time the subresource is accessed.

If imageView is not VK_NULL_HANDLE, imageLayout must be VK_IMAGE_LAYOUT_SHADING_RATE_OPTIMAL_NV or VK_IMAGE_LAYOUT_GENERAL.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

imageView must be a valid VkImageView handle

imageLayout must be a valid VkImageLayout value

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics operations

Both of commandBuffer, and imageView must have been created, allocated, or retrieved from the same VkDevice

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkImageLayout, VkImageView

cmdSetCoarseSampleOrderNV :: CommandBuffer -> CoarseSampleOrderTypeNV -> Vector CoarseSampleOrderCustomNV -> IO ()

vkCmdSetCoarseSampleOrderNV - Set sample order for coarse fragments on a command buffer

Parameters

commandBuffer is the command buffer into which the command will be recorded.

sampleOrderType specifies the mechanism used to order coverage samples in fragments larger than one pixel.

customSampleOrderCount specifies the number of custom sample orderings to use when ordering coverage samples.

pCustomSampleOrders is a pointer to an array of VkCoarseSampleOrderCustomNV structures, each of which specifies the coverage sample order for a single combination of fragment area and coverage sample count.

Description

If sampleOrderType is VK_COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV, the coverage sample order used for any combination of fragment area and coverage sample count not enumerated in pCustomSampleOrders will be identical to that used for VK_COARSE_SAMPLE_ORDER_TYPE_DEFAULT_NV.

Valid Usage

If sampleOrderType is not VK_COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV, customSamplerOrderCount must be 0

The array pCustomSampleOrders must not contain two structures with matching values for both the shadingRate and sampleCount members.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

sampleOrderType must be a valid VkCoarseSampleOrderTypeNV value

If customSampleOrderCount is not 0, pCustomSampleOrders must be a valid pointer to an array of customSampleOrderCount valid VkCoarseSampleOrderCustomNV structures

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCoarseSampleOrderCustomNV, VkCoarseSampleOrderTypeNV, VkCommandBuffer

cmdSetViewportShadingRatePaletteNV :: CommandBuffer -> Word32 -> Vector ShadingRatePaletteNV -> IO ()

vkCmdSetViewportShadingRatePaletteNV - Set shading rate image palettes on a command buffer

Parameters

commandBuffer is the command buffer into which the command will be recorded.

firstViewport is the index of the first viewport whose shading rate palette is updated by the command.

viewportCount is the number of viewports whose shading rate palettes are updated by the command.

pShadingRatePalettes is a pointer to an array of VkShadingRatePaletteNV structures defining the palette for each viewport.

Valid Usage

The shading rate image feature must be enabled.

The bound graphics pipeline must have been created with the VK_DYNAMIC_STATE_VIEWPORT_SHADING_RATE_PALETTE_NV dynamic state enabled

firstViewport must be less than VkPhysicalDeviceLimits::maxViewports

The sum of firstViewport and viewportCount must be between 1 and VkPhysicalDeviceLimits::maxViewports, inclusive

If the multiple viewports feature is not enabled, firstViewport must be 0

If the multiple viewports feature is not enabled, viewportCount must be 1

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pShadingRatePalettes must be a valid pointer to an array of viewportCount valid VkShadingRatePaletteNV structures

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics operations

viewportCount must be greater than 0

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 |                 |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkShadingRatePaletteNV

pattern NV_SHADING_RATE_IMAGE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NV_SHADING_RATE_IMAGE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PIPELINE_VIEWPORT_SHADING_RATE_IMAGE_STATE_CREATE_INFO_NV :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADING_RATE_IMAGE_FEATURES_NV :: a ~ StructureType => a

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADING_RATE_IMAGE_PROPERTIES_NV :: a ~ StructureType => a

pattern IMAGE_LAYOUT_SHADING_RATE_OPTIMAL_NV :: a ~ ImageLayout => a

pattern DYNAMIC_STATE_VIEWPORT_SHADING_RATE_PALETTE_NV :: a ~ DynamicState => a

pattern ACCESS_SHADING_RATE_IMAGE_READ_BIT_NV :: a ~ AccessFlagBits => a

pattern IMAGE_USAGE_SHADING_RATE_IMAGE_BIT_NV :: a ~ ImageUsageFlagBits => a

pattern PIPELINE_STAGE_SHADING_RATE_IMAGE_BIT_NV :: a ~ PipelineStageFlagBits => a

pattern STRUCTURE_TYPE_PIPELINE_VIEWPORT_COARSE_SAMPLE_ORDER_STATE_CREATE_INFO_NV :: a ~ StructureType => a

pattern DYNAMIC_STATE_VIEWPORT_COARSE_SAMPLE_ORDER_NV :: a ~ DynamicState => a