Graphics.Vulkan.Extensions.VK_NV_viewport_swizzle

Documentation

type PipelineViewportSwizzleStateCreateFlagsNV = VkPipelineViewportSwizzleStateCreateFlagsNV

VkPipelineViewportSwizzleStateCreateFlagsNV - Reserved for future use

Description

VkPipelineViewportSwizzleStateCreateFlagsNV is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

VkPipelineViewportSwizzleStateCreateInfoNV

withCStructPipelineViewportSwizzleStateCreateInfoNV :: PipelineViewportSwizzleStateCreateInfoNV -> (VkPipelineViewportSwizzleStateCreateInfoNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkPipelineViewportSwizzleStateCreateInfoNV and marshal a PipelineViewportSwizzleStateCreateInfoNV into it. The VkPipelineViewportSwizzleStateCreateInfoNV is only valid inside the provided computation and must not be returned out of it.

fromCStructPipelineViewportSwizzleStateCreateInfoNV :: VkPipelineViewportSwizzleStateCreateInfoNV -> IO PipelineViewportSwizzleStateCreateInfoNV

A function to read a VkPipelineViewportSwizzleStateCreateInfoNV and all additional structures in the pointer chain into a PipelineViewportSwizzleStateCreateInfoNV.

data PipelineViewportSwizzleStateCreateInfoNV

VkPipelineViewportSwizzleStateCreateInfoNV - Structure specifying swizzle applied to primitive clip coordinates

Valid Usage (Implicit)

See Also

VkPipelineViewportSwizzleStateCreateFlagsNV, VkStructureType, VkViewportSwizzleNV

Constructors

PipelineViewportSwizzleStateCreateInfoNV

Fields

next :: Maybe SomeVkStruct

flags :: PipelineViewportSwizzleStateCreateFlagsNV

viewportSwizzles :: Vector ViewportSwizzleNV

type ViewportCoordinateSwizzleNV = VkViewportCoordinateSwizzleNV

VkViewportCoordinateSwizzleNV - Specify how a viewport coordinate is swizzled

Description

These values are described in detail in Viewport Swizzle.

See Also

VkViewportSwizzleNV

pattern VIEWPORT_COORDINATE_SWIZZLE_POSITIVE_X_NV :: a ~ ViewportCoordinateSwizzleNV => a

pattern VIEWPORT_COORDINATE_SWIZZLE_NEGATIVE_X_NV :: a ~ ViewportCoordinateSwizzleNV => a

pattern VIEWPORT_COORDINATE_SWIZZLE_POSITIVE_Y_NV :: a ~ ViewportCoordinateSwizzleNV => a

pattern VIEWPORT_COORDINATE_SWIZZLE_NEGATIVE_Y_NV :: a ~ ViewportCoordinateSwizzleNV => a

pattern VIEWPORT_COORDINATE_SWIZZLE_POSITIVE_Z_NV :: a ~ ViewportCoordinateSwizzleNV => a

pattern VIEWPORT_COORDINATE_SWIZZLE_NEGATIVE_Z_NV :: a ~ ViewportCoordinateSwizzleNV => a

pattern VIEWPORT_COORDINATE_SWIZZLE_POSITIVE_W_NV :: a ~ ViewportCoordinateSwizzleNV => a

pattern VIEWPORT_COORDINATE_SWIZZLE_NEGATIVE_W_NV :: a ~ ViewportCoordinateSwizzleNV => a

withCStructViewportSwizzleNV :: ViewportSwizzleNV -> (VkViewportSwizzleNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkViewportSwizzleNV and marshal a ViewportSwizzleNV into it. The VkViewportSwizzleNV is only valid inside the provided computation and must not be returned out of it.

fromCStructViewportSwizzleNV :: VkViewportSwizzleNV -> IO ViewportSwizzleNV

A function to read a VkViewportSwizzleNV and all additional structures in the pointer chain into a ViewportSwizzleNV.

data ViewportSwizzleNV

VkViewportSwizzleNV - Structure specifying a viewport swizzle

Valid Usage (Implicit)

See Also

VkPipelineViewportSwizzleStateCreateInfoNV, VkViewportCoordinateSwizzleNV

Constructors

ViewportSwizzleNV

Fields

x :: ViewportCoordinateSwizzleNV

y :: ViewportCoordinateSwizzleNV

z :: ViewportCoordinateSwizzleNV

w :: ViewportCoordinateSwizzleNV

pattern NV_VIEWPORT_SWIZZLE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NV_VIEWPORT_SWIZZLE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_PIPELINE_VIEWPORT_SWIZZLE_STATE_CREATE_INFO_NV :: a ~ StructureType => a