Graphics.Vulkan.Extensions.VK_NV_win32_keyed_mutex

Documentation

withCStructWin32KeyedMutexAcquireReleaseInfoNV :: Win32KeyedMutexAcquireReleaseInfoNV -> (VkWin32KeyedMutexAcquireReleaseInfoNV -> IO a) -> IO a

A function to temporarily allocate memory for a VkWin32KeyedMutexAcquireReleaseInfoNV and marshal a Win32KeyedMutexAcquireReleaseInfoNV into it. The VkWin32KeyedMutexAcquireReleaseInfoNV is only valid inside the provided computation and must not be returned out of it.

fromCStructWin32KeyedMutexAcquireReleaseInfoNV :: VkWin32KeyedMutexAcquireReleaseInfoNV -> IO Win32KeyedMutexAcquireReleaseInfoNV

A function to read a VkWin32KeyedMutexAcquireReleaseInfoNV and all additional structures in the pointer chain into a Win32KeyedMutexAcquireReleaseInfoNV.

data Win32KeyedMutexAcquireReleaseInfoNV

VkWin32KeyedMutexAcquireReleaseInfoNV - use Windows keyex mutex mechanism to synchronize work

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_WIN32_KEYED_MUTEX_ACQUIRE_RELEASE_INFO_NV

If acquireCount is not 0, pAcquireSyncs must be a valid pointer to an array of acquireCount valid VkDeviceMemory handles

If acquireCount is not 0, pAcquireKeys must be a valid pointer to an array of acquireCount uint64_t values

If acquireCount is not 0, pAcquireTimeoutMilliseconds must be a valid pointer to an array of acquireCount uint32_t values

If releaseCount is not 0, pReleaseSyncs must be a valid pointer to an array of releaseCount valid VkDeviceMemory handles

If releaseCount is not 0, pReleaseKeys must be a valid pointer to an array of releaseCount uint64_t values

Both of the elements of pAcquireSyncs, and the elements of pReleaseSyncs that are valid handles must have been created, allocated, or retrieved from the same VkDevice

See Also

VkDeviceMemory, VkStructureType

Constructors

Win32KeyedMutexAcquireReleaseInfoNV

Fields

next :: Maybe SomeVkStruct

acquireSyncs :: Vector DeviceMemory

acquireKeys :: Vector Word64

acquireTimeoutMilliseconds :: Vector Word32

releaseSyncs :: Vector DeviceMemory

releaseKeys :: Vector Word64

pattern NV_WIN32_KEYED_MUTEX_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NV_WIN32_KEYED_MUTEX_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_WIN32_KEYED_MUTEX_ACQUIRE_RELEASE_INFO_NV :: a ~ StructureType => a