Graphics.Vulkan.Marshal.SomeVkStruct

Documentation

class ToCStruct marshalled c | marshalled -> c, c -> marshalled where

A class for types which can be marshalled into a C style structure.

Methods

withCStruct :: marshalled -> (c -> IO a) -> IO a

Allocates a C type structure and all dependencies and passes it to a continuation. The space is deallocated when this continuation returns and the C type structure must not be returned out of it.

class FromCStruct marshalled c | marshalled -> c, c -> marshalled where

A class for converting C type structures to the marshalled types

Methods

fromCStruct :: c -> IO marshalled

Read a C type structure and dependencies

data SomeVkStruct where

A wrapper for holding any Vulkan struct, this is used to implement structure pNext chains.

Constructors

SomeVkStruct :: (ToCStruct a b, Storable b, Show a, Eq a, Typeable a, HasNext a) => a -> SomeVkStruct

class HasNext a where

A class for types which can be present in a pNext chain

Methods

getNext :: a -> Maybe SomeVkStruct

fromSomeVkStruct :: Typeable a => SomeVkStruct -> Maybe a

Convert a SomeVkStruct to a structure of a known type if possible.

fromSomeVkStructChain :: Typeable a => SomeVkStruct -> Maybe a

Search the whole pointer chain for a structure of a particular type and return that if possible

withSomeVkStruct :: SomeVkStruct -> (Ptr () -> IO a) -> IO a

Allocate space for the value contained in a SomeVkStruct and use that in continuation.

withCStructPtr :: (Storable c, ToCStruct a c) => a -> (Ptr c -> IO b) -> IO b

fromCStructPtr :: (Storable c, FromCStruct a c) => Ptr c -> IO a

fromCStructPtrElem :: (Storable c, FromCStruct a c) => Ptr c -> Int -> IO a

peekVkStruct :: Ptr SomeVkStruct -> IO SomeVkStruct

Read the sType member of a Vulkan struct and marshal the struct into a SomeVkStruct

Make sure that you only pass this a pointer to a Vulkan struct with a sType member at offset 0 otherwise the behaviour is undefined.

Throws an InvalidArgument IOException if given a pointer to a struct with an unrecognised sType member.

Throws an InvalidArgument IOException if given a pointer to a struct which can't be marshalled (those containing union types)