Graphics.Vulkan.Marshal.Utils

Documentation

withVec :: forall a b d. Storable b => (a -> (b -> IO d) -> IO d) -> Vector a -> (Ptr b -> IO d) -> IO d

withArray :: forall a b d. (a -> (b -> IO d) -> IO d) -> Vector a -> (Vector b -> IO d) -> IO d

withSizedArray :: forall a b d n. (a -> (b -> IO d) -> IO d) -> Vector n a -> (Vector n b -> IO d) -> IO d

byteStringToSizedVector :: forall n v. (KnownNat n, Vector v Word8) => ByteString -> Vector v n Word8

Convert a bytestring to a sized vector. If the bytestring is too long it will be truncated. If it is too short it will be zero padded

byteStringToNullTerminatedSizedVector :: forall n v. (KnownNat n, 1 <= n, Vector v CChar) => ByteString -> Vector v n CChar

Convert a bytestring to a null terminated sized vector. If the bytestring is too long it will be truncated.

padSized

Arguments

:: (KnownNat n, Vector v a)

=> a
The value with which to pad if the given vector is too short

-> v a
The vector to pad or truncate

-> Vector v n a

Pad or truncate a vector so that it has the required size

padVector :: Vector v a => a -> Int -> v a -> v a

Make sure a vector is at least a certain length

packCStringElemOff :: Ptr (Ptr CChar) -> Int -> IO ByteString

peekVkStruct :: Ptr SomeVkStruct -> IO SomeVkStruct