Vulkan.CStruct.Extends

Documentation

data BaseOutStructure

VkBaseOutStructure - Base structure for a read-only pointer chain

Description

BaseOutStructure can be used to facilitate iterating through a structure pointer chain that returns data back to the application.

See Also

BaseOutStructure, StructureType

Constructors

BaseOutStructure

Fields

sType :: StructureType

sType is the structure type of the structure being iterated through.

next :: Ptr BaseOutStructure

pNext is NULL or a pointer to the next structure in a structure chain.

data BaseInStructure

VkBaseInStructure - Base structure for a read-only pointer chain

Description

BaseInStructure can be used to facilitate iterating through a read-only structure pointer chain.

See Also

BaseInStructure, StructureType

Constructors

BaseInStructure

Fields

sType :: StructureType

sType is the structure type of the structure being iterated through.

next :: Ptr BaseInStructure

pNext is NULL or a pointer to the next structure in a structure chain.

type family Extends (a :: [Type] -> Type) (b :: Type) :: Constraint where ...

Equations

Extends AccelerationStructureBuildGeometryInfoKHR DeferredOperationInfoKHR = ()

Extends AndroidHardwareBufferPropertiesANDROID AndroidHardwareBufferFormatPropertiesANDROID = ()

Extends AttachmentDescription2 AttachmentDescriptionStencilLayout = ()

Extends AttachmentReference2 AttachmentReferenceStencilLayout = ()

Extends BindBufferMemoryInfo BindBufferMemoryDeviceGroupInfo = ()

Extends BindImageMemoryInfo BindImageMemoryDeviceGroupInfo = ()

Extends BindImageMemoryInfo BindImageMemorySwapchainInfoKHR = ()

Extends BindImageMemoryInfo BindImagePlaneMemoryInfo = ()

Extends BindSparseInfo DeviceGroupBindSparseInfo = ()

Extends BindSparseInfo TimelineSemaphoreSubmitInfo = ()

Extends BufferCreateInfo DedicatedAllocationBufferCreateInfoNV = ()

Extends BufferCreateInfo ExternalMemoryBufferCreateInfo = ()

Extends BufferCreateInfo BufferOpaqueCaptureAddressCreateInfo = ()

Extends BufferCreateInfo BufferDeviceAddressCreateInfoEXT = ()

Extends CommandBufferBeginInfo DeviceGroupCommandBufferBeginInfo = ()

Extends CommandBufferInheritanceInfo CommandBufferInheritanceConditionalRenderingInfoEXT = ()

Extends CommandBufferInheritanceInfo CommandBufferInheritanceRenderPassTransformInfoQCOM = ()

Extends ComputePipelineCreateInfo PipelineCreationFeedbackCreateInfoEXT = ()

Extends ComputePipelineCreateInfo PipelineCompilerControlCreateInfoAMD = ()

Extends CopyAccelerationStructureInfoKHR DeferredOperationInfoKHR = ()

Extends CopyAccelerationStructureToMemoryInfoKHR DeferredOperationInfoKHR = ()

Extends CopyMemoryToAccelerationStructureInfoKHR DeferredOperationInfoKHR = ()

Extends DescriptorPoolCreateInfo DescriptorPoolInlineUniformBlockCreateInfoEXT = ()

Extends DescriptorSetAllocateInfo DescriptorSetVariableDescriptorCountAllocateInfo = ()

Extends DescriptorSetLayoutCreateInfo DescriptorSetLayoutBindingFlagsCreateInfo = ()

Extends DescriptorSetLayoutSupport DescriptorSetVariableDescriptorCountLayoutSupport = ()

Extends DeviceCreateInfo PhysicalDeviceDeviceGeneratedCommandsFeaturesNV = ()

Extends DeviceCreateInfo DevicePrivateDataCreateInfoEXT = ()

Extends DeviceCreateInfo PhysicalDevicePrivateDataFeaturesEXT = ()

Extends DeviceCreateInfo (PhysicalDeviceFeatures2 '[]) = ()

Extends DeviceCreateInfo PhysicalDeviceVariablePointersFeatures = ()

Extends DeviceCreateInfo PhysicalDeviceMultiviewFeatures = ()

Extends DeviceCreateInfo DeviceGroupDeviceCreateInfo = ()

Extends DeviceCreateInfo PhysicalDevice16BitStorageFeatures = ()

Extends DeviceCreateInfo PhysicalDeviceShaderSubgroupExtendedTypesFeatures = ()

Extends DeviceCreateInfo PhysicalDeviceSamplerYcbcrConversionFeatures = ()

Extends DeviceCreateInfo PhysicalDeviceProtectedMemoryFeatures = ()

Extends DeviceCreateInfo PhysicalDeviceBlendOperationAdvancedFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceInlineUniformBlockFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceShaderDrawParametersFeatures = ()

Extends DeviceCreateInfo PhysicalDeviceShaderFloat16Int8Features = ()

Extends DeviceCreateInfo PhysicalDeviceHostQueryResetFeatures = ()

Extends DeviceCreateInfo PhysicalDeviceDeviceMemoryReportFeaturesEXT = ()

Extends DeviceCreateInfo DeviceDeviceMemoryReportCreateInfoEXT = ()

Extends DeviceCreateInfo PhysicalDeviceDescriptorIndexingFeatures = ()

Extends DeviceCreateInfo PhysicalDeviceTimelineSemaphoreFeatures = ()

Extends DeviceCreateInfo PhysicalDevice8BitStorageFeatures = ()

Extends DeviceCreateInfo PhysicalDeviceConditionalRenderingFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceVulkanMemoryModelFeatures = ()

Extends DeviceCreateInfo PhysicalDeviceShaderAtomicInt64Features = ()

Extends DeviceCreateInfo PhysicalDeviceShaderAtomicFloatFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceVertexAttributeDivisorFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceASTCDecodeFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceTransformFeedbackFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceRepresentativeFragmentTestFeaturesNV = ()

Extends DeviceCreateInfo PhysicalDeviceExclusiveScissorFeaturesNV = ()

Extends DeviceCreateInfo PhysicalDeviceCornerSampledImageFeaturesNV = ()

Extends DeviceCreateInfo PhysicalDeviceComputeShaderDerivativesFeaturesNV = ()

Extends DeviceCreateInfo PhysicalDeviceFragmentShaderBarycentricFeaturesNV = ()

Extends DeviceCreateInfo PhysicalDeviceShaderImageFootprintFeaturesNV = ()

Extends DeviceCreateInfo PhysicalDeviceDedicatedAllocationImageAliasingFeaturesNV = ()

Extends DeviceCreateInfo PhysicalDeviceShadingRateImageFeaturesNV = ()

Extends DeviceCreateInfo PhysicalDeviceMeshShaderFeaturesNV = ()

Extends DeviceCreateInfo PhysicalDeviceRayTracingFeaturesKHR = ()

Extends DeviceCreateInfo DeviceMemoryOverallocationCreateInfoAMD = ()

Extends DeviceCreateInfo PhysicalDeviceFragmentDensityMapFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceFragmentDensityMap2FeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceScalarBlockLayoutFeatures = ()

Extends DeviceCreateInfo PhysicalDeviceUniformBufferStandardLayoutFeatures = ()

Extends DeviceCreateInfo PhysicalDeviceDepthClipEnableFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceMemoryPriorityFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceBufferDeviceAddressFeatures = ()

Extends DeviceCreateInfo PhysicalDeviceBufferDeviceAddressFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceImagelessFramebufferFeatures = ()

Extends DeviceCreateInfo PhysicalDeviceTextureCompressionASTCHDRFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceCooperativeMatrixFeaturesNV = ()

Extends DeviceCreateInfo PhysicalDeviceYcbcrImageArraysFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDevicePerformanceQueryFeaturesKHR = ()

Extends DeviceCreateInfo PhysicalDeviceCoverageReductionModeFeaturesNV = ()

Extends DeviceCreateInfo PhysicalDeviceShaderIntegerFunctions2FeaturesINTEL = ()

Extends DeviceCreateInfo PhysicalDeviceShaderClockFeaturesKHR = ()

Extends DeviceCreateInfo PhysicalDeviceIndexTypeUint8FeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceShaderSMBuiltinsFeaturesNV = ()

Extends DeviceCreateInfo PhysicalDeviceFragmentShaderInterlockFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceSeparateDepthStencilLayoutsFeatures = ()

Extends DeviceCreateInfo PhysicalDevicePipelineExecutablePropertiesFeaturesKHR = ()

Extends DeviceCreateInfo PhysicalDeviceShaderDemoteToHelperInvocationFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceTexelBufferAlignmentFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceSubgroupSizeControlFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceLineRasterizationFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDevicePipelineCreationCacheControlFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceVulkan11Features = ()

Extends DeviceCreateInfo PhysicalDeviceVulkan12Features = ()

Extends DeviceCreateInfo PhysicalDeviceCoherentMemoryFeaturesAMD = ()

Extends DeviceCreateInfo PhysicalDeviceCustomBorderColorFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceExtendedDynamicStateFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceDiagnosticsConfigFeaturesNV = ()

Extends DeviceCreateInfo DeviceDiagnosticsConfigCreateInfoNV = ()

Extends DeviceCreateInfo PhysicalDeviceRobustness2FeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceImageRobustnessFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDevicePortabilitySubsetFeaturesKHR = ()

Extends DeviceCreateInfo PhysicalDevice4444FormatsFeaturesEXT = ()

Extends DeviceCreateInfo PhysicalDeviceShaderImageAtomicInt64FeaturesEXT = ()

Extends DeviceQueueCreateInfo DeviceQueueGlobalPriorityCreateInfoEXT = ()

Extends FenceCreateInfo ExportFenceCreateInfo = ()

Extends FenceCreateInfo ExportFenceWin32HandleInfoKHR = ()

Extends FormatProperties2 DrmFormatModifierPropertiesListEXT = ()

Extends FramebufferCreateInfo FramebufferAttachmentsCreateInfo = ()

Extends GraphicsPipelineCreateInfo GraphicsPipelineShaderGroupsCreateInfoNV = ()

Extends GraphicsPipelineCreateInfo PipelineDiscardRectangleStateCreateInfoEXT = ()

Extends GraphicsPipelineCreateInfo PipelineRepresentativeFragmentTestStateCreateInfoNV = ()

Extends GraphicsPipelineCreateInfo PipelineCreationFeedbackCreateInfoEXT = ()

Extends GraphicsPipelineCreateInfo PipelineCompilerControlCreateInfoAMD = ()

Extends ImageCreateInfo DedicatedAllocationImageCreateInfoNV = ()

Extends ImageCreateInfo ExternalMemoryImageCreateInfoNV = ()

Extends ImageCreateInfo ExternalMemoryImageCreateInfo = ()

Extends ImageCreateInfo ImageSwapchainCreateInfoKHR = ()

Extends ImageCreateInfo ImageFormatListCreateInfo = ()

Extends ImageCreateInfo ExternalFormatANDROID = ()

Extends ImageCreateInfo ImageDrmFormatModifierListCreateInfoEXT = ()

Extends ImageCreateInfo ImageDrmFormatModifierExplicitCreateInfoEXT = ()

Extends ImageCreateInfo ImageStencilUsageCreateInfo = ()

Extends ImageFormatProperties2 ExternalImageFormatProperties = ()

Extends ImageFormatProperties2 SamplerYcbcrConversionImageFormatProperties = ()

Extends ImageFormatProperties2 TextureLODGatherFormatPropertiesAMD = ()

Extends ImageFormatProperties2 AndroidHardwareBufferUsageANDROID = ()

Extends ImageFormatProperties2 FilterCubicImageViewImageFormatPropertiesEXT = ()

Extends ImageMemoryBarrier SampleLocationsInfoEXT = ()

Extends ImageMemoryRequirementsInfo2 ImagePlaneMemoryRequirementsInfo = ()

Extends ImageViewCreateInfo ImageViewUsageCreateInfo = ()

Extends ImageViewCreateInfo SamplerYcbcrConversionInfo = ()

Extends ImageViewCreateInfo ImageViewASTCDecodeModeEXT = ()

Extends InstanceCreateInfo DebugReportCallbackCreateInfoEXT = ()

Extends InstanceCreateInfo ValidationFlagsEXT = ()

Extends InstanceCreateInfo ValidationFeaturesEXT = ()

Extends InstanceCreateInfo DebugUtilsMessengerCreateInfoEXT = ()

Extends MemoryAllocateInfo DedicatedAllocationMemoryAllocateInfoNV = ()

Extends MemoryAllocateInfo ExportMemoryAllocateInfoNV = ()

Extends MemoryAllocateInfo ImportMemoryWin32HandleInfoNV = ()

Extends MemoryAllocateInfo ExportMemoryWin32HandleInfoNV = ()

Extends MemoryAllocateInfo ExportMemoryAllocateInfo = ()

Extends MemoryAllocateInfo ImportMemoryWin32HandleInfoKHR = ()

Extends MemoryAllocateInfo ExportMemoryWin32HandleInfoKHR = ()

Extends MemoryAllocateInfo ImportMemoryFdInfoKHR = ()

Extends MemoryAllocateInfo MemoryAllocateFlagsInfo = ()

Extends MemoryAllocateInfo MemoryDedicatedAllocateInfo = ()

Extends MemoryAllocateInfo ImportMemoryHostPointerInfoEXT = ()

Extends MemoryAllocateInfo ImportAndroidHardwareBufferInfoANDROID = ()

Extends MemoryAllocateInfo MemoryPriorityAllocateInfoEXT = ()

Extends MemoryAllocateInfo MemoryOpaqueCaptureAddressAllocateInfo = ()

Extends MemoryRequirements2 MemoryDedicatedRequirements = ()

Extends PhysicalDeviceExternalSemaphoreInfo SemaphoreTypeCreateInfo = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceDeviceGeneratedCommandsFeaturesNV = ()

Extends PhysicalDeviceFeatures2 PhysicalDevicePrivateDataFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceVariablePointersFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceMultiviewFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDevice16BitStorageFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceShaderSubgroupExtendedTypesFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceSamplerYcbcrConversionFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceProtectedMemoryFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceBlendOperationAdvancedFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceInlineUniformBlockFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceShaderDrawParametersFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceShaderFloat16Int8Features = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceHostQueryResetFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceDeviceMemoryReportFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceDescriptorIndexingFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceTimelineSemaphoreFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDevice8BitStorageFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceConditionalRenderingFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceVulkanMemoryModelFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceShaderAtomicInt64Features = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceShaderAtomicFloatFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceVertexAttributeDivisorFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceASTCDecodeFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceTransformFeedbackFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceRepresentativeFragmentTestFeaturesNV = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceExclusiveScissorFeaturesNV = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceCornerSampledImageFeaturesNV = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceComputeShaderDerivativesFeaturesNV = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceFragmentShaderBarycentricFeaturesNV = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceShaderImageFootprintFeaturesNV = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceDedicatedAllocationImageAliasingFeaturesNV = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceShadingRateImageFeaturesNV = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceMeshShaderFeaturesNV = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceRayTracingFeaturesKHR = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceFragmentDensityMapFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceFragmentDensityMap2FeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceScalarBlockLayoutFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceUniformBufferStandardLayoutFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceDepthClipEnableFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceMemoryPriorityFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceBufferDeviceAddressFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceBufferDeviceAddressFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceImagelessFramebufferFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceTextureCompressionASTCHDRFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceCooperativeMatrixFeaturesNV = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceYcbcrImageArraysFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDevicePerformanceQueryFeaturesKHR = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceCoverageReductionModeFeaturesNV = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceShaderIntegerFunctions2FeaturesINTEL = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceShaderClockFeaturesKHR = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceIndexTypeUint8FeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceShaderSMBuiltinsFeaturesNV = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceFragmentShaderInterlockFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceSeparateDepthStencilLayoutsFeatures = ()

Extends PhysicalDeviceFeatures2 PhysicalDevicePipelineExecutablePropertiesFeaturesKHR = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceShaderDemoteToHelperInvocationFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceTexelBufferAlignmentFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceSubgroupSizeControlFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceLineRasterizationFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDevicePipelineCreationCacheControlFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceVulkan11Features = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceVulkan12Features = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceCoherentMemoryFeaturesAMD = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceCustomBorderColorFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceExtendedDynamicStateFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceDiagnosticsConfigFeaturesNV = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceRobustness2FeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceImageRobustnessFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDevicePortabilitySubsetFeaturesKHR = ()

Extends PhysicalDeviceFeatures2 PhysicalDevice4444FormatsFeaturesEXT = ()

Extends PhysicalDeviceFeatures2 PhysicalDeviceShaderImageAtomicInt64FeaturesEXT = ()

Extends PhysicalDeviceImageFormatInfo2 PhysicalDeviceExternalImageFormatInfo = ()

Extends PhysicalDeviceImageFormatInfo2 ImageFormatListCreateInfo = ()

Extends PhysicalDeviceImageFormatInfo2 PhysicalDeviceImageDrmFormatModifierInfoEXT = ()

Extends PhysicalDeviceImageFormatInfo2 ImageStencilUsageCreateInfo = ()

Extends PhysicalDeviceImageFormatInfo2 PhysicalDeviceImageViewImageFormatInfoEXT = ()

Extends PhysicalDeviceMemoryProperties2 PhysicalDeviceMemoryBudgetPropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceDeviceGeneratedCommandsPropertiesNV = ()

Extends PhysicalDeviceProperties2 PhysicalDevicePushDescriptorPropertiesKHR = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceDriverProperties = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceIDProperties = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceMultiviewProperties = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceDiscardRectanglePropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceMultiviewPerViewAttributesPropertiesNVX = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceSubgroupProperties = ()

Extends PhysicalDeviceProperties2 PhysicalDevicePointClippingProperties = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceProtectedMemoryProperties = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceSamplerFilterMinmaxProperties = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceSampleLocationsPropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceBlendOperationAdvancedPropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceInlineUniformBlockPropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceMaintenance3Properties = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceFloatControlsProperties = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceExternalMemoryHostPropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceConservativeRasterizationPropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceShaderCorePropertiesAMD = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceShaderCoreProperties2AMD = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceDescriptorIndexingProperties = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceTimelineSemaphoreProperties = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceVertexAttributeDivisorPropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDevicePCIBusInfoPropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceDepthStencilResolveProperties = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceTransformFeedbackPropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceShadingRateImagePropertiesNV = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceMeshShaderPropertiesNV = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceRayTracingPropertiesKHR = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceRayTracingPropertiesNV = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceFragmentDensityMapPropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceFragmentDensityMap2PropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceCooperativeMatrixPropertiesNV = ()

Extends PhysicalDeviceProperties2 PhysicalDevicePerformanceQueryPropertiesKHR = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceShaderSMBuiltinsPropertiesNV = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceTexelBufferAlignmentPropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceSubgroupSizeControlPropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceLineRasterizationPropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceVulkan11Properties = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceVulkan12Properties = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceCustomBorderColorPropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDeviceRobustness2PropertiesEXT = ()

Extends PhysicalDeviceProperties2 PhysicalDevicePortabilitySubsetPropertiesKHR = ()

Extends PhysicalDeviceSurfaceInfo2KHR SurfaceFullScreenExclusiveInfoEXT = ()

Extends PhysicalDeviceSurfaceInfo2KHR SurfaceFullScreenExclusiveWin32InfoEXT = ()

Extends PipelineColorBlendStateCreateInfo PipelineColorBlendAdvancedStateCreateInfoEXT = ()

Extends PipelineMultisampleStateCreateInfo PipelineCoverageToColorStateCreateInfoNV = ()

Extends PipelineMultisampleStateCreateInfo PipelineSampleLocationsStateCreateInfoEXT = ()

Extends PipelineMultisampleStateCreateInfo PipelineCoverageModulationStateCreateInfoNV = ()

Extends PipelineMultisampleStateCreateInfo PipelineCoverageReductionStateCreateInfoNV = ()

Extends PipelineRasterizationStateCreateInfo PipelineRasterizationStateRasterizationOrderAMD = ()

Extends PipelineRasterizationStateCreateInfo PipelineRasterizationConservativeStateCreateInfoEXT = ()

Extends PipelineRasterizationStateCreateInfo PipelineRasterizationStateStreamCreateInfoEXT = ()

Extends PipelineRasterizationStateCreateInfo PipelineRasterizationDepthClipStateCreateInfoEXT = ()

Extends PipelineRasterizationStateCreateInfo PipelineRasterizationLineStateCreateInfoEXT = ()

Extends PipelineShaderStageCreateInfo PipelineShaderStageRequiredSubgroupSizeCreateInfoEXT = ()

Extends PipelineTessellationStateCreateInfo PipelineTessellationDomainOriginStateCreateInfo = ()

Extends PipelineVertexInputStateCreateInfo PipelineVertexInputDivisorStateCreateInfoEXT = ()

Extends PipelineViewportStateCreateInfo PipelineViewportWScalingStateCreateInfoNV = ()

Extends PipelineViewportStateCreateInfo PipelineViewportSwizzleStateCreateInfoNV = ()

Extends PipelineViewportStateCreateInfo PipelineViewportExclusiveScissorStateCreateInfoNV = ()

Extends PipelineViewportStateCreateInfo PipelineViewportShadingRateImageStateCreateInfoNV = ()

Extends PipelineViewportStateCreateInfo PipelineViewportCoarseSampleOrderStateCreateInfoNV = ()

Extends PresentInfoKHR DisplayPresentInfoKHR = ()

Extends PresentInfoKHR PresentRegionsKHR = ()

Extends PresentInfoKHR DeviceGroupPresentInfoKHR = ()

Extends PresentInfoKHR PresentTimesInfoGOOGLE = ()

Extends PresentInfoKHR PresentFrameTokenGGP = ()

Extends QueryPoolCreateInfo QueryPoolPerformanceCreateInfoKHR = ()

Extends QueryPoolCreateInfo QueryPoolPerformanceQueryCreateInfoINTEL = ()

Extends QueueFamilyProperties2 QueueFamilyCheckpointPropertiesNV = ()

Extends RayTracingPipelineCreateInfoKHR PipelineCreationFeedbackCreateInfoEXT = ()

Extends RayTracingPipelineCreateInfoKHR DeferredOperationInfoKHR = ()

Extends RayTracingPipelineCreateInfoNV PipelineCreationFeedbackCreateInfoEXT = ()

Extends RenderPassBeginInfo DeviceGroupRenderPassBeginInfo = ()

Extends RenderPassBeginInfo RenderPassSampleLocationsBeginInfoEXT = ()

Extends RenderPassBeginInfo RenderPassAttachmentBeginInfo = ()

Extends RenderPassBeginInfo RenderPassTransformBeginInfoQCOM = ()

Extends RenderPassCreateInfo RenderPassMultiviewCreateInfo = ()

Extends RenderPassCreateInfo RenderPassInputAttachmentAspectCreateInfo = ()

Extends RenderPassCreateInfo RenderPassFragmentDensityMapCreateInfoEXT = ()

Extends RenderPassCreateInfo2 RenderPassFragmentDensityMapCreateInfoEXT = ()

Extends SamplerCreateInfo SamplerYcbcrConversionInfo = ()

Extends SamplerCreateInfo SamplerReductionModeCreateInfo = ()

Extends SamplerCreateInfo SamplerCustomBorderColorCreateInfoEXT = ()

Extends SamplerYcbcrConversionCreateInfo ExternalFormatANDROID = ()

Extends SemaphoreCreateInfo ExportSemaphoreCreateInfo = ()

Extends SemaphoreCreateInfo ExportSemaphoreWin32HandleInfoKHR = ()

Extends SemaphoreCreateInfo SemaphoreTypeCreateInfo = ()

Extends ShaderModuleCreateInfo ShaderModuleValidationCacheCreateInfoEXT = ()

Extends SubmitInfo Win32KeyedMutexAcquireReleaseInfoNV = ()

Extends SubmitInfo Win32KeyedMutexAcquireReleaseInfoKHR = ()

Extends SubmitInfo D3D12FenceSubmitInfoKHR = ()

Extends SubmitInfo DeviceGroupSubmitInfo = ()

Extends SubmitInfo ProtectedSubmitInfo = ()

Extends SubmitInfo TimelineSemaphoreSubmitInfo = ()

Extends SubmitInfo PerformanceQuerySubmitInfoKHR = ()

Extends SubpassDescription2 SubpassDescriptionDepthStencilResolve = ()

Extends SurfaceCapabilities2KHR DisplayNativeHdrSurfaceCapabilitiesAMD = ()

Extends SurfaceCapabilities2KHR SharedPresentSurfaceCapabilitiesKHR = ()

Extends SurfaceCapabilities2KHR SurfaceProtectedCapabilitiesKHR = ()

Extends SurfaceCapabilities2KHR SurfaceCapabilitiesFullScreenExclusiveEXT = ()

Extends SwapchainCreateInfoKHR SwapchainCounterCreateInfoEXT = ()

Extends SwapchainCreateInfoKHR DeviceGroupSwapchainCreateInfoKHR = ()

Extends SwapchainCreateInfoKHR SwapchainDisplayNativeHdrCreateInfoAMD = ()

Extends SwapchainCreateInfoKHR ImageFormatListCreateInfo = ()

Extends SwapchainCreateInfoKHR SurfaceFullScreenExclusiveInfoEXT = ()

Extends SwapchainCreateInfoKHR SurfaceFullScreenExclusiveWin32InfoEXT = ()

Extends WriteDescriptorSet WriteDescriptorSetInlineUniformBlockEXT = ()

Extends WriteDescriptorSet WriteDescriptorSetAccelerationStructureKHR = ()

Extends a b = TypeError ((ShowType a :<>: Text " is not extended by ") :<>: ShowType b)

class PeekChain es where

Methods

peekChain :: Ptr (Chain es) -> IO (Chain es)

class PokeChain es where

Methods

withChain :: Chain es -> (Ptr (Chain es) -> IO a) -> IO a

withZeroChain :: (Ptr (Chain es) -> IO a) -> IO a

type family Chain (xs :: [a]) = (r :: a) | r -> xs where ...

Equations

Chain '[] = ()

Chain (x ': xs) = (x, Chain xs)

type family Extendss (p :: [Type] -> Type) (xs :: [Type]) :: Constraint where ...

Equations

Extendss p '[] = ()

Extendss p (x ': xs) = (Extends p x, Extendss p xs)

data SomeStruct (a :: [Type] -> Type) where

Constructors

SomeStruct :: forall a es. (Extendss a es, PokeChain es, Show (Chain es)) => a es -> SomeStruct a

extendSomeStruct :: (Extensible a, Extends a e, ToCStruct e, Show e) => e -> SomeStruct a -> SomeStruct a

Add an extension to the beginning of the struct chain

This can be used to optionally extend structs based on some condition (for example, an extension or layer being available)

withSomeStruct :: forall a b. SomeStruct a -> (forall es. (Extendss a es, PokeChain es, Show (Chain es)) => a es -> b) -> b

Consume a SomeStruct value

withSomeCStruct :: forall a b. (forall es. (Extendss a es, PokeChain es) => ToCStruct (a es)) => SomeStruct a -> (forall es. (Extendss a es, PokeChain es) => Ptr (a es) -> IO b) -> IO b

Write the C representation of some extended a and use the pointer, the pointer must not be returned from the continuation.

peekSomeCStruct :: forall a. (Extensible a, forall es. (Extendss a es, PeekChain es) => FromCStruct (a es)) => Ptr (SomeStruct a) -> IO (SomeStruct a)

Given a pointer to a struct with an unknown chain, peek the struct and its chain.

pokeSomeCStruct

Arguments

:: (forall es. (Extendss a es, PokeChain es) => ToCStruct (a es))

=> Ptr (SomeStruct a)
Pointer to some memory at least the size of the head of the struct chain.

-> SomeStruct a
The struct to poke

-> IO b
Computation to run while the poked tail is valid

-> IO b

Given some memory for the head of the chain, allocate and poke the tail and run an action.

forgetExtensions :: Ptr (a es) -> Ptr (SomeStruct a)

Forget which extensions a pointed-to struct has by casting the pointer

class Extensible (a :: [Type] -> Type) where

Methods

extensibleType :: StructureType

getNext :: a es -> Chain es

setNext :: a ds -> Chain es -> a es

extends :: forall e b proxy. Typeable e => proxy e -> (Extends a e => b) -> Maybe b

pattern (::&) :: Extensible a => a es' -> Chain es -> a es infix 6

A pattern synonym to separate the head of a struct chain from the tail, use in conjunction with :& to extract several members.

Head{..} ::& () <- returningNoTail a b c
-- Equivalent to
Head{..} <- returningNoTail @'[] a b c

Head{..} ::& Foo{..} :& Bar{..} :& () <- returningWithTail a b c

myFun (Head{..} :&& Foo{..} :& ())

pattern (:&) :: e -> Chain es -> Chain (e ': es) infixr 7

View the head and tail of a Chain, see ::&

Equivalent to (,)