Vulkan.Core10.APIConstants

Documentation

pattern LOD_CLAMP_NONE :: Float

type LUID_SIZE_KHR = LUID_SIZE

type QUEUE_FAMILY_EXTERNAL_KHR = QUEUE_FAMILY_EXTERNAL

type MAX_DEVICE_GROUP_SIZE_KHR = MAX_DEVICE_GROUP_SIZE

type MAX_DRIVER_NAME_SIZE_KHR = MAX_DRIVER_NAME_SIZE

type MAX_DRIVER_INFO_SIZE_KHR = MAX_DRIVER_INFO_SIZE

type SHADER_UNUSED_NV = SHADER_UNUSED_KHR

type MAX_PHYSICAL_DEVICE_NAME_SIZE = 256

pattern MAX_PHYSICAL_DEVICE_NAME_SIZE :: forall a. Integral a => a

type UUID_SIZE = 16

pattern UUID_SIZE :: forall a. Integral a => a

type LUID_SIZE = 8

pattern LUID_SIZE :: forall a. Integral a => a

type MAX_EXTENSION_NAME_SIZE = 256

pattern MAX_EXTENSION_NAME_SIZE :: forall a. Integral a => a

type MAX_DESCRIPTION_SIZE = 256

pattern MAX_DESCRIPTION_SIZE :: forall a. Integral a => a

type MAX_MEMORY_TYPES = 32

pattern MAX_MEMORY_TYPES :: forall a. Integral a => a

type MAX_MEMORY_HEAPS = 16

pattern MAX_MEMORY_HEAPS :: forall a. Integral a => a

type REMAINING_MIP_LEVELS = 4294967295

pattern REMAINING_MIP_LEVELS :: Word32

type REMAINING_ARRAY_LAYERS = 4294967295

pattern REMAINING_ARRAY_LAYERS :: Word32

type WHOLE_SIZE = 18446744073709551615

pattern WHOLE_SIZE :: Word64

type ATTACHMENT_UNUSED = 4294967295

pattern ATTACHMENT_UNUSED :: Word32

type QUEUE_FAMILY_IGNORED = 4294967295

pattern QUEUE_FAMILY_IGNORED :: Word32

type QUEUE_FAMILY_EXTERNAL = 4294967294

pattern QUEUE_FAMILY_EXTERNAL :: Word32

type QUEUE_FAMILY_FOREIGN_EXT = 4294967293

pattern QUEUE_FAMILY_FOREIGN_EXT :: Word32

type SUBPASS_EXTERNAL = 4294967295

pattern SUBPASS_EXTERNAL :: Word32

type MAX_DEVICE_GROUP_SIZE = 32

pattern MAX_DEVICE_GROUP_SIZE :: forall a. Integral a => a

type MAX_DRIVER_NAME_SIZE = 256

pattern MAX_DRIVER_NAME_SIZE :: forall a. Integral a => a

type MAX_DRIVER_INFO_SIZE = 256

pattern MAX_DRIVER_INFO_SIZE :: forall a. Integral a => a

type SHADER_UNUSED_KHR = 4294967295

pattern SHADER_UNUSED_KHR :: Word32

pattern NULL_HANDLE :: IsHandle a => a

VK_NULL_HANDLE - Reserved non-valid object handle

See Also

No cross-references are available

class (Eq a, Zero a) => IsHandle a

A class for things which can be created with NULL_HANDLE.

class HasObjectType a where

Methods

objectTypeAndHandle :: a -> (ObjectType, Word64)

newtype Bool32

VkBool32 - Vulkan boolean type

Description

TRUE represents a boolean True (integer 1) value, and FALSE a boolean False (integer 0) value.

All values returned from a Vulkan implementation in a Bool32 will be either TRUE or FALSE.

Applications must not pass any other values than TRUE or FALSE into a Vulkan implementation where a Bool32 is expected.

See Also

AccelerationStructureBuildGeometryInfoKHR, AccelerationStructureCreateGeometryTypeInfoKHR, AccelerationStructureGeometryInstancesDataKHR, CommandBufferInheritanceConditionalRenderingInfoEXT, CommandBufferInheritanceInfo, DedicatedAllocationBufferCreateInfoNV, DedicatedAllocationImageCreateInfoNV, DescriptorSetLayoutSupport, DisplayNativeHdrSurfaceCapabilitiesAMD, DisplayPresentInfoKHR, DisplayPropertiesKHR, FilterCubicImageViewImageFormatPropertiesEXT, IndirectCommandsLayoutTokenNV, MemoryDedicatedRequirements, PerformanceOverrideInfoINTEL, PerformanceValueDataINTEL, PhysicalDevice16BitStorageFeatures, PhysicalDevice4444FormatsFeaturesEXT, PhysicalDevice8BitStorageFeatures, PhysicalDeviceASTCDecodeFeaturesEXT, PhysicalDeviceBlendOperationAdvancedFeaturesEXT, PhysicalDeviceBlendOperationAdvancedPropertiesEXT, PhysicalDeviceBufferDeviceAddressFeatures, PhysicalDeviceBufferDeviceAddressFeaturesEXT, PhysicalDeviceCoherentMemoryFeaturesAMD, PhysicalDeviceComputeShaderDerivativesFeaturesNV, PhysicalDeviceConditionalRenderingFeaturesEXT, PhysicalDeviceConservativeRasterizationPropertiesEXT, PhysicalDeviceCooperativeMatrixFeaturesNV, PhysicalDeviceCornerSampledImageFeaturesNV, PhysicalDeviceCoverageReductionModeFeaturesNV, PhysicalDeviceCustomBorderColorFeaturesEXT, PhysicalDeviceDedicatedAllocationImageAliasingFeaturesNV, PhysicalDeviceDepthClipEnableFeaturesEXT, PhysicalDeviceDepthStencilResolveProperties, PhysicalDeviceDescriptorIndexingFeatures, PhysicalDeviceDescriptorIndexingProperties, PhysicalDeviceDeviceGeneratedCommandsFeaturesNV, PhysicalDeviceDiagnosticsConfigFeaturesNV, PhysicalDeviceExclusiveScissorFeaturesNV, PhysicalDeviceExtendedDynamicStateFeaturesEXT, PhysicalDeviceFeatures, PhysicalDeviceFloatControlsProperties, PhysicalDeviceFragmentDensityMap2FeaturesEXT, PhysicalDeviceFragmentDensityMap2PropertiesEXT, PhysicalDeviceFragmentDensityMapFeaturesEXT, PhysicalDeviceFragmentDensityMapPropertiesEXT, PhysicalDeviceFragmentShaderBarycentricFeaturesNV, PhysicalDeviceFragmentShaderInterlockFeaturesEXT, PhysicalDeviceGroupProperties, PhysicalDeviceHostQueryResetFeatures, PhysicalDeviceIDProperties, PhysicalDeviceImageRobustnessFeaturesEXT, PhysicalDeviceImagelessFramebufferFeatures, PhysicalDeviceIndexTypeUint8FeaturesEXT, PhysicalDeviceInlineUniformBlockFeaturesEXT, PhysicalDeviceLimits, PhysicalDeviceLineRasterizationFeaturesEXT, PhysicalDeviceMemoryPriorityFeaturesEXT, PhysicalDeviceMeshShaderFeaturesNV, PhysicalDeviceMultiviewFeatures, PhysicalDeviceMultiviewPerViewAttributesPropertiesNVX, PhysicalDevicePerformanceQueryFeaturesKHR, PhysicalDevicePerformanceQueryPropertiesKHR, PhysicalDevicePipelineCreationCacheControlFeaturesEXT, PhysicalDevicePipelineExecutablePropertiesFeaturesKHR, PhysicalDevicePrivateDataFeaturesEXT, PhysicalDeviceProtectedMemoryFeatures, PhysicalDeviceProtectedMemoryProperties, PhysicalDeviceRayTracingFeaturesKHR, PhysicalDeviceRepresentativeFragmentTestFeaturesNV, PhysicalDeviceRobustness2FeaturesEXT, PhysicalDeviceSampleLocationsPropertiesEXT, PhysicalDeviceSamplerFilterMinmaxProperties, PhysicalDeviceSamplerYcbcrConversionFeatures, PhysicalDeviceScalarBlockLayoutFeatures, PhysicalDeviceSeparateDepthStencilLayoutsFeatures, PhysicalDeviceShaderAtomicFloatFeaturesEXT, PhysicalDeviceShaderAtomicInt64Features, PhysicalDeviceShaderClockFeaturesKHR, PhysicalDeviceShaderDemoteToHelperInvocationFeaturesEXT, PhysicalDeviceShaderDrawParametersFeatures, PhysicalDeviceShaderFloat16Int8Features, PhysicalDeviceShaderImageFootprintFeaturesNV, PhysicalDeviceShaderIntegerFunctions2FeaturesINTEL, PhysicalDeviceShaderSMBuiltinsFeaturesNV, PhysicalDeviceShaderSubgroupExtendedTypesFeatures, PhysicalDeviceShadingRateImageFeaturesNV, PhysicalDeviceSparseProperties, PhysicalDeviceSubgroupProperties, PhysicalDeviceSubgroupSizeControlFeaturesEXT, PhysicalDeviceTexelBufferAlignmentFeaturesEXT, PhysicalDeviceTexelBufferAlignmentPropertiesEXT, PhysicalDeviceTextureCompressionASTCHDRFeaturesEXT, PhysicalDeviceTimelineSemaphoreFeatures, PhysicalDeviceTransformFeedbackFeaturesEXT, PhysicalDeviceTransformFeedbackPropertiesEXT, PhysicalDeviceUniformBufferStandardLayoutFeatures, PhysicalDeviceVariablePointersFeatures, PhysicalDeviceVertexAttributeDivisorFeaturesEXT, PhysicalDeviceVulkan11Features, PhysicalDeviceVulkan11Properties, PhysicalDeviceVulkan12Features, PhysicalDeviceVulkan12Properties, PhysicalDeviceVulkanMemoryModelFeatures, PhysicalDeviceYcbcrImageArraysFeaturesEXT, PipelineColorBlendAdvancedStateCreateInfoEXT, PipelineColorBlendAttachmentState, PipelineColorBlendStateCreateInfo, PipelineCoverageModulationStateCreateInfoNV, PipelineCoverageToColorStateCreateInfoNV, PipelineDepthStencilStateCreateInfo, PipelineExecutableInternalRepresentationKHR, PipelineExecutableStatisticValueKHR, PipelineInputAssemblyStateCreateInfo, PipelineMultisampleStateCreateInfo, PipelineRasterizationDepthClipStateCreateInfoEXT, PipelineRasterizationLineStateCreateInfoEXT, PipelineRasterizationStateCreateInfo, PipelineRepresentativeFragmentTestStateCreateInfoNV, PipelineSampleLocationsStateCreateInfoEXT, PipelineViewportShadingRateImageStateCreateInfoNV, PipelineViewportWScalingStateCreateInfoNV, ProtectedSubmitInfo, SamplerCreateInfo, SamplerYcbcrConversionCreateInfo, SurfaceCapabilitiesFullScreenExclusiveEXT, SurfaceProtectedCapabilitiesKHR, SwapchainCreateInfoKHR, SwapchainDisplayNativeHdrCreateInfoAMD, TextureLODGatherFormatPropertiesAMD, cmdBuildAccelerationStructureNV, cmdExecuteGeneratedCommandsNV, cmdSetDepthBoundsTestEnableEXT, cmdSetDepthTestEnableEXT, cmdSetDepthWriteEnableEXT, cmdSetStencilTestEnableEXT, getPhysicalDeviceSurfaceSupportKHR, setLocalDimmingAMD, waitForFences

Constructors

Bool32 Int32

Bundled Patterns

pattern FALSE :: Bool32

pattern TRUE :: Bool32

newtype PipelineCacheHeaderVersion

VkPipelineCacheHeaderVersion - Encode pipeline cache version

See Also

createPipelineCache, getPipelineCacheData

Constructors

PipelineCacheHeaderVersion Int32

Bundled Patterns

pattern PIPELINE_CACHE_HEADER_VERSION_ONE :: PipelineCacheHeaderVersion
PIPELINE_CACHE_HEADER_VERSION_ONE specifies version one of the pipeline cache.