Vulkan.Core10.Buffer

Documentation

createBuffer

Arguments

:: forall a io. (Extendss BufferCreateInfo a, PokeChain a, MonadIO io)

=> Device
device is the logical device that creates the buffer object.

-> BufferCreateInfo a
pCreateInfo is a pointer to a BufferCreateInfo structure containing parameters affecting creation of the buffer.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io Buffer

vkCreateBuffer - Create a new buffer object

Valid Usage

If the flags member of pCreateInfo includes BUFFER_CREATE_SPARSE_BINDING_BIT, creating this Buffer must not cause the total required sparse memory for all currently valid sparse resources on the device to exceed PhysicalDeviceLimits::sparseAddressSpaceSize

Valid Usage (Implicit)

device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid BufferCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pBuffer must be a valid pointer to a Buffer handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_INVALID_OPAQUE_CAPTURE_ADDRESS_KHR

See Also

AllocationCallbacks, Buffer, BufferCreateInfo, Device

withBuffer :: forall a io r. (Extendss BufferCreateInfo a, PokeChain a, MonadIO io) => Device -> BufferCreateInfo a -> Maybe AllocationCallbacks -> (io Buffer -> (Buffer -> io ()) -> r) -> r

A convenience wrapper to make a compatible pair of calls to createBuffer and destroyBuffer

To ensure that destroyBuffer is always called: pass bracket (or the allocate function from your favourite resource management library) as the last argument. To just extract the pair pass (,) as the last argument.

destroyBuffer

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that destroys the buffer.

-> Buffer
buffer is the buffer to destroy.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io ()

vkDestroyBuffer - Destroy a buffer object

Valid Usage

All submitted commands that refer to buffer, either directly or via a BufferView, must have completed execution

If AllocationCallbacks were provided when buffer was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when buffer was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid Device handle

If buffer is not NULL_HANDLE, buffer must be a valid Buffer handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If buffer is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to buffer must be externally synchronized

See Also

AllocationCallbacks, Buffer, Device

data BufferCreateInfo (es :: [Type])

VkBufferCreateInfo - Structure specifying the parameters of a newly created buffer object

Valid Usage

size must be greater than 0

If sharingMode is SHARING_MODE_CONCURRENT, pQueueFamilyIndices must be a valid pointer to an array of queueFamilyIndexCount uint32_t values

If sharingMode is SHARING_MODE_CONCURRENT, queueFamilyIndexCount must be greater than 1

If sharingMode is SHARING_MODE_CONCURRENT, each element of pQueueFamilyIndices must be unique and must be less than pQueueFamilyPropertyCount returned by either getPhysicalDeviceQueueFamilyProperties or getPhysicalDeviceQueueFamilyProperties2 for the physicalDevice that was used to create device

If the sparse bindings feature is not enabled, flags must not contain BUFFER_CREATE_SPARSE_BINDING_BIT

If the sparse buffer residency feature is not enabled, flags must not contain BUFFER_CREATE_SPARSE_RESIDENCY_BIT

If the sparse aliased residency feature is not enabled, flags must not contain BUFFER_CREATE_SPARSE_ALIASED_BIT

If flags contains BUFFER_CREATE_SPARSE_RESIDENCY_BIT or BUFFER_CREATE_SPARSE_ALIASED_BIT, it must also contain BUFFER_CREATE_SPARSE_BINDING_BIT

If the pNext chain includes a ExternalMemoryBufferCreateInfo structure, its handleTypes member must only contain bits that are also in ExternalBufferProperties::externalMemoryProperties.compatibleHandleTypes, as returned by getPhysicalDeviceExternalBufferProperties with pExternalBufferInfo->handleType equal to any one of the handle types specified in ExternalMemoryBufferCreateInfo::handleTypes

If the protected memory feature is not enabled, flags must not contain BUFFER_CREATE_PROTECTED_BIT

If any of the bits BUFFER_CREATE_SPARSE_BINDING_BIT, BUFFER_CREATE_SPARSE_RESIDENCY_BIT, or BUFFER_CREATE_SPARSE_ALIASED_BIT are set, BUFFER_CREATE_PROTECTED_BIT must not also be set

If the pNext chain includes a DedicatedAllocationBufferCreateInfoNV structure, and the dedicatedAllocation member of the chained structure is TRUE, then flags must not include BUFFER_CREATE_SPARSE_BINDING_BIT, BUFFER_CREATE_SPARSE_RESIDENCY_BIT, or BUFFER_CREATE_SPARSE_ALIASED_BIT

If BufferDeviceAddressCreateInfoEXT::deviceAddress is not zero, flags must include BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT

If BufferOpaqueCaptureAddressCreateInfo::opaqueCaptureAddress is not zero, flags must include BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT

If flags includes BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT, the bufferDeviceAddressCaptureReplay or ::bufferDeviceAddressCaptureReplay feature must be enabled

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_BUFFER_CREATE_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of BufferDeviceAddressCreateInfoEXT, BufferOpaqueCaptureAddressCreateInfo, DedicatedAllocationBufferCreateInfoNV, or ExternalMemoryBufferCreateInfo

The sType value of each struct in the pNext chain must be unique

flags must be a valid combination of BufferCreateFlagBits values

usage must be a valid combination of BufferUsageFlagBits values

usage must not be 0

sharingMode must be a valid SharingMode value

See Also

BufferCreateFlags, BufferUsageFlags, DeviceSize, SharingMode, StructureType, createBuffer

Constructors

BufferCreateInfo

Fields

next :: Chain es

pNext is NULL or a pointer to a structure extending this structure.

flags :: BufferCreateFlags

flags is a bitmask of BufferCreateFlagBits specifying additional parameters of the buffer.

size :: DeviceSize

size is the size in bytes of the buffer to be created.

usage :: BufferUsageFlags

usage is a bitmask of BufferUsageFlagBits specifying allowed usages of the buffer.

sharingMode :: SharingMode

sharingMode is a SharingMode value specifying the sharing mode of the buffer when it will be accessed by multiple queue families.

queueFamilyIndices :: Vector Word32

pQueueFamilyIndices is a list of queue families that will access this buffer (ignored if sharingMode is not SHARING_MODE_CONCURRENT).

newtype Buffer

VkBuffer - Opaque handle to a buffer object

See Also

AccelerationStructureCreateInfoKHR, BindBufferMemoryInfo, BufferDeviceAddressInfo, BufferMemoryBarrier, BufferMemoryRequirementsInfo2, BufferViewCreateInfo, ConditionalRenderingBeginInfoEXT, CopyBufferInfo2KHR, CopyBufferToImageInfo2KHR, CopyImageToBufferInfo2KHR, DedicatedAllocationMemoryAllocateInfoNV, DescriptorBufferInfo, GeneratedCommandsInfoNV, GeometryAABBNV, GeometryTrianglesNV, IndirectCommandsStreamNV, MemoryDedicatedAllocateInfo, SparseBufferMemoryBindInfo, bindBufferMemory, cmdBeginTransformFeedbackEXT, cmdBindIndexBuffer, cmdBindTransformFeedbackBuffersEXT, cmdBindVertexBuffers, cmdBindVertexBuffers2EXT, cmdBuildAccelerationStructureNV, cmdCopyBuffer, cmdCopyBufferToImage, cmdCopyImageToBuffer, cmdCopyQueryPoolResults, cmdDispatchIndirect, cmdDrawIndexedIndirect, cmdDrawIndexedIndirectCount, cmdDrawIndexedIndirectCountAMD, cmdDrawIndexedIndirectCountKHR, cmdDrawIndirect, cmdDrawIndirectByteCountEXT, cmdDrawIndirectCount, cmdDrawIndirectCountAMD, cmdDrawIndirectCountKHR, cmdDrawMeshTasksIndirectCountNV, cmdDrawMeshTasksIndirectNV, cmdEndTransformFeedbackEXT, cmdFillBuffer, cmdTraceRaysNV, cmdUpdateBuffer, cmdWriteBufferMarkerAMD, createBuffer, destroyBuffer, getBufferMemoryRequirements

Constructors

Buffer Word64

newtype SharingMode

VkSharingMode - Buffer and image sharing modes

Description

Note

SHARING_MODE_CONCURRENT may result in lower performance access to the buffer or image than SHARING_MODE_EXCLUSIVE.

Ranges of buffers and image subresources of image objects created using SHARING_MODE_EXCLUSIVE must only be accessed by queues in the queue family that has ownership of the resource. Upon creation, such resources are not owned by any queue family; ownership is implicitly acquired upon first use within a queue. Once a resource using SHARING_MODE_EXCLUSIVE is owned by some queue family, the application must perform a queue family ownership transfer to make the memory contents of a range or image subresource accessible to a different queue family.

Note

Images still require a layout transition from IMAGE_LAYOUT_UNDEFINED or IMAGE_LAYOUT_PREINITIALIZED before being used on the first queue.

A queue family can take ownership of an image subresource or buffer range of a resource created with SHARING_MODE_EXCLUSIVE, without an ownership transfer, in the same way as for a resource that was just created; however, taking ownership in this way has the effect that the contents of the image subresource or buffer range are undefined.

Ranges of buffers and image subresources of image objects created using SHARING_MODE_CONCURRENT must only be accessed by queues from the queue families specified through the queueFamilyIndexCount and pQueueFamilyIndices members of the corresponding create info structures.

See Also

BufferCreateInfo, ImageCreateInfo, PhysicalDeviceImageDrmFormatModifierInfoEXT, SwapchainCreateInfoKHR

Constructors

SharingMode Int32

Bundled Patterns

pattern SHARING_MODE_EXCLUSIVE :: SharingMode
SHARING_MODE_EXCLUSIVE specifies that access to any range or image subresource of the object will be exclusive to a single queue family at a time.

pattern SHARING_MODE_CONCURRENT :: SharingMode
SHARING_MODE_CONCURRENT specifies that concurrent access to any range or image subresource of the object from multiple queue families is supported.

newtype BufferUsageFlagBits

VkBufferUsageFlagBits - Bitmask specifying allowed usage of a buffer

See Also

BufferUsageFlags

Constructors

BufferUsageFlagBits Flags

Bundled Patterns

pattern BUFFER_USAGE_TRANSFER_SRC_BIT :: BufferUsageFlagBits
BUFFER_USAGE_TRANSFER_SRC_BIT specifies that the buffer can be used as the source of a transfer command (see the definition of https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#synchronization-pipeline-stages-transfer).

pattern BUFFER_USAGE_TRANSFER_DST_BIT :: BufferUsageFlagBits
BUFFER_USAGE_TRANSFER_DST_BIT specifies that the buffer can be used as the destination of a transfer command.

pattern BUFFER_USAGE_UNIFORM_TEXEL_BUFFER_BIT :: BufferUsageFlagBits
BUFFER_USAGE_UNIFORM_TEXEL_BUFFER_BIT specifies that the buffer can be used to create a BufferView suitable for occupying a DescriptorSet slot of type DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER.

pattern BUFFER_USAGE_STORAGE_TEXEL_BUFFER_BIT :: BufferUsageFlagBits
BUFFER_USAGE_STORAGE_TEXEL_BUFFER_BIT specifies that the buffer can be used to create a BufferView suitable for occupying a DescriptorSet slot of type DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER.

pattern BUFFER_USAGE_UNIFORM_BUFFER_BIT :: BufferUsageFlagBits
BUFFER_USAGE_UNIFORM_BUFFER_BIT specifies that the buffer can be used in a DescriptorBufferInfo suitable for occupying a DescriptorSet slot either of type DESCRIPTOR_TYPE_UNIFORM_BUFFER or DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC.

pattern BUFFER_USAGE_STORAGE_BUFFER_BIT :: BufferUsageFlagBits
BUFFER_USAGE_STORAGE_BUFFER_BIT specifies that the buffer can be used in a DescriptorBufferInfo suitable for occupying a DescriptorSet slot either of type DESCRIPTOR_TYPE_STORAGE_BUFFER or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC.

pattern BUFFER_USAGE_INDEX_BUFFER_BIT :: BufferUsageFlagBits
BUFFER_USAGE_INDEX_BUFFER_BIT specifies that the buffer is suitable for passing as the buffer parameter to cmdBindIndexBuffer.

pattern BUFFER_USAGE_VERTEX_BUFFER_BIT :: BufferUsageFlagBits
BUFFER_USAGE_VERTEX_BUFFER_BIT specifies that the buffer is suitable for passing as an element of the pBuffers array to cmdBindVertexBuffers.

pattern BUFFER_USAGE_INDIRECT_BUFFER_BIT :: BufferUsageFlagBits
BUFFER_USAGE_INDIRECT_BUFFER_BIT specifies that the buffer is suitable for passing as the buffer parameter to cmdDrawIndirect, cmdDrawIndexedIndirect, cmdDrawMeshTasksIndirectNV, cmdDrawMeshTasksIndirectCountNV, or cmdDispatchIndirect. It is also suitable for passing as the buffer member of IndirectCommandsStreamNV, or sequencesCountBuffer or sequencesIndexBuffer or preprocessedBuffer member of GeneratedCommandsInfoNV

pattern BUFFER_USAGE_SHADER_BINDING_TABLE_BIT_KHR :: BufferUsageFlagBits
BUFFER_USAGE_SHADER_BINDING_TABLE_BIT_KHR specifies that the buffer is suitable for use as a Shader Binding Table.

pattern BUFFER_USAGE_ACCELERATION_STRUCTURE_STORAGE_BIT_KHR :: BufferUsageFlagBits
BUFFER_USAGE_ACCELERATION_STRUCTURE_STORAGE_BIT_KHR specifies that the buffer is suitable for storage space for a AccelerationStructureKHR.

pattern BUFFER_USAGE_ACCELERATION_STRUCTURE_BUILD_INPUT_READ_ONLY_BIT_KHR :: BufferUsageFlagBits
BUFFER_USAGE_ACCELERATION_STRUCTURE_BUILD_INPUT_READ_ONLY_BIT_KHR specifies that the buffer is suitable for use as a read-only input to an acceleration structure build.

pattern BUFFER_USAGE_CONDITIONAL_RENDERING_BIT_EXT :: BufferUsageFlagBits
BUFFER_USAGE_CONDITIONAL_RENDERING_BIT_EXT specifies that the buffer is suitable for passing as the buffer parameter to cmdBeginConditionalRenderingEXT.

pattern BUFFER_USAGE_TRANSFORM_FEEDBACK_COUNTER_BUFFER_BIT_EXT :: BufferUsageFlagBits
BUFFER_USAGE_TRANSFORM_FEEDBACK_COUNTER_BUFFER_BIT_EXT specifies that the buffer is suitable for using as a counter buffer with cmdBeginTransformFeedbackEXT and cmdEndTransformFeedbackEXT.

pattern BUFFER_USAGE_TRANSFORM_FEEDBACK_BUFFER_BIT_EXT :: BufferUsageFlagBits
BUFFER_USAGE_TRANSFORM_FEEDBACK_BUFFER_BIT_EXT specifies that the buffer is suitable for using for binding as a transform feedback buffer with cmdBindTransformFeedbackBuffersEXT.

pattern BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT :: BufferUsageFlagBits
BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT specifies that the buffer can be used to retrieve a buffer device address via getBufferDeviceAddress and use that address to access the buffer’s memory from a shader.

type BufferUsageFlags = BufferUsageFlagBits

newtype BufferCreateFlagBits

VkBufferCreateFlagBits - Bitmask specifying additional parameters of a buffer

Description

See Sparse Resource Features and Physical Device Features for details of the sparse memory features supported on a device.

See Also

BufferCreateFlags

Constructors

BufferCreateFlagBits Flags

Bundled Patterns

pattern BUFFER_CREATE_SPARSE_BINDING_BIT :: BufferCreateFlagBits
BUFFER_CREATE_SPARSE_BINDING_BIT specifies that the buffer will be backed using sparse memory binding.

pattern BUFFER_CREATE_SPARSE_RESIDENCY_BIT :: BufferCreateFlagBits
BUFFER_CREATE_SPARSE_RESIDENCY_BIT specifies that the buffer can be partially backed using sparse memory binding. Buffers created with this flag must also be created with the BUFFER_CREATE_SPARSE_BINDING_BIT flag.

pattern BUFFER_CREATE_SPARSE_ALIASED_BIT :: BufferCreateFlagBits
BUFFER_CREATE_SPARSE_ALIASED_BIT specifies that the buffer will be backed using sparse memory binding with memory ranges that might also simultaneously be backing another buffer (or another portion of the same buffer). Buffers created with this flag must also be created with the BUFFER_CREATE_SPARSE_BINDING_BIT flag.

pattern BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT :: BufferCreateFlagBits
BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT specifies that the buffer’s address can be saved and reused on a subsequent run (e.g. for trace capture and replay), see BufferOpaqueCaptureAddressCreateInfo for more detail.

pattern BUFFER_CREATE_PROTECTED_BIT :: BufferCreateFlagBits
BUFFER_CREATE_PROTECTED_BIT specifies that the buffer is a protected buffer.

type BufferCreateFlags = BufferCreateFlagBits