Vulkan.Core10.CommandBuffer

Documentation

allocateCommandBuffers

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that owns the command pool.

-> CommandBufferAllocateInfo
pAllocateInfo is a pointer to a CommandBufferAllocateInfo structure describing parameters of the allocation.

-> io ("commandBuffers" ::: Vector CommandBuffer)

vkAllocateCommandBuffers - Allocate command buffers from an existing command pool

Description

allocateCommandBuffers can be used to create multiple command buffers. If the creation of any of those command buffers fails, the implementation must destroy all successfully created command buffer objects from this command, set all entries of the pCommandBuffers array to NULL and return the error.

When command buffers are first allocated, they are in the initial state.

Valid Usage (Implicit)

device must be a valid Device handle

pAllocateInfo must be a valid pointer to a valid CommandBufferAllocateInfo structure

pCommandBuffers must be a valid pointer to an array of pAllocateInfo->commandBufferCount CommandBuffer handles

pAllocateInfo->commandBufferCount must be greater than 0

Host Synchronization

Host access to pAllocateInfo->commandPool must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

CommandBuffer, CommandBufferAllocateInfo, Device

withCommandBuffers :: forall io r. MonadIO io => Device -> CommandBufferAllocateInfo -> (io (Vector CommandBuffer) -> (Vector CommandBuffer -> io ()) -> r) -> r

A convenience wrapper to make a compatible pair of calls to allocateCommandBuffers and freeCommandBuffers

To ensure that freeCommandBuffers is always called: pass bracket (or the allocate function from your favourite resource management library) as the first argument. To just extract the pair pass (,) as the first argument.

freeCommandBuffers

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that owns the command pool.

-> CommandPool
commandPool is the command pool from which the command buffers were allocated.

-> ("commandBuffers" ::: Vector CommandBuffer)
pCommandBuffers is a pointer to an array of handles of command buffers to free.

-> io ()

vkFreeCommandBuffers - Free command buffers

Description

Any primary command buffer that is in the recording or executable state and has any element of pCommandBuffers recorded into it, becomes invalid.

Valid Usage

All elements of pCommandBuffers must not be in the pending state

pCommandBuffers must be a valid pointer to an array of commandBufferCount CommandBuffer handles, each element of which must either be a valid handle or NULL

Valid Usage (Implicit)

device must be a valid Device handle

commandPool must be a valid CommandPool handle

commandBufferCount must be greater than 0

commandPool must have been created, allocated, or retrieved from device

Each element of pCommandBuffers that is a valid handle must have been created, allocated, or retrieved from commandPool

Host Synchronization

Host access to commandPool must be externally synchronized

Host access to each member of pCommandBuffers must be externally synchronized

See Also

CommandBuffer, CommandPool, Device

beginCommandBuffer

Arguments

:: forall a io. (Extendss CommandBufferBeginInfo a, PokeChain a, MonadIO io)

=> CommandBuffer
commandBuffer is the handle of the command buffer which is to be put in the recording state.

-> CommandBufferBeginInfo a
pBeginInfo points to a CommandBufferBeginInfo structure defining additional information about how the command buffer begins recording.

-> io ()

vkBeginCommandBuffer - Start recording a command buffer

Valid Usage

commandBuffer must not be in the recording or pending state

If commandBuffer was allocated from a CommandPool which did not have the COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT flag set, commandBuffer must be in the initial state

If commandBuffer is a secondary command buffer, the pInheritanceInfo member of pBeginInfo must be a valid CommandBufferInheritanceInfo structure

If commandBuffer is a secondary command buffer and either the occlusionQueryEnable member of the pInheritanceInfo member of pBeginInfo is FALSE, or the precise occlusion queries feature is not enabled, the queryFlags member of the pInheritanceInfo member pBeginInfo must not contain QUERY_CONTROL_PRECISE_BIT

If commandBuffer is a primary command buffer, then pBeginInfo->flags must not set both the COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT and the COMMAND_BUFFER_USAGE_SIMULTANEOUS_USE_BIT flags

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pBeginInfo must be a valid pointer to a valid CommandBufferBeginInfo structure

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

CommandBuffer, CommandBufferBeginInfo

useCommandBuffer :: forall a io r. (Extendss CommandBufferBeginInfo a, PokeChain a, MonadIO io) => CommandBuffer -> CommandBufferBeginInfo a -> io r -> io r

This function will call the supplied action between calls to beginCommandBuffer and endCommandBuffer

Note that endCommandBuffer is *not* called if an exception is thrown by the inner action.

endCommandBuffer

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer to complete recording.

-> io ()

vkEndCommandBuffer - Finish recording a command buffer

Description

If there was an error during recording, the application will be notified by an unsuccessful return code returned by endCommandBuffer. If the application wishes to further use the command buffer, the command buffer must be reset. The command buffer must have been in the recording state, and is moved to the executable state.

Valid Usage

commandBuffer must be in the recording state

If commandBuffer is a primary command buffer, there must not be an active render pass instance

All queries made active during the recording of commandBuffer must have been made inactive

Conditional rendering must not be active

If commandBuffer is a secondary command buffer, there must not be an outstanding cmdBeginDebugUtilsLabelEXT command recorded to commandBuffer that has not previously been ended by a call to cmdEndDebugUtilsLabelEXT

If commandBuffer is a secondary command buffer, there must not be an outstanding cmdDebugMarkerBeginEXT command recorded to commandBuffer that has not previously been ended by a call to cmdDebugMarkerEndEXT

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

CommandBuffer

resetCommandBuffer

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer to reset. The command buffer can be in any state other than pending, and is moved into the initial state.

-> CommandBufferResetFlags
flags is a bitmask of CommandBufferResetFlagBits controlling the reset operation.

-> io ()

vkResetCommandBuffer - Reset a command buffer to the initial state

Description

Any primary command buffer that is in the recording or executable state and has commandBuffer recorded into it, becomes invalid.

Valid Usage

commandBuffer must not be in the pending state

commandBuffer must have been allocated from a pool that was created with the COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

flags must be a valid combination of CommandBufferResetFlagBits values

Host Synchronization

Host access to commandBuffer must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_DEVICE_MEMORY

See Also

CommandBuffer, CommandBufferResetFlags

data CommandBufferAllocateInfo

VkCommandBufferAllocateInfo - Structure specifying the allocation parameters for command buffer object

Valid Usage (Implicit)

See Also

CommandBufferLevel, CommandPool, StructureType, allocateCommandBuffers

Constructors

CommandBufferAllocateInfo

Fields

commandPool :: CommandPool

commandPool is the command pool from which the command buffers are allocated.

commandPool must be a valid CommandPool handle

level :: CommandBufferLevel

level is a CommandBufferLevel value specifying the command buffer level.

level must be a valid CommandBufferLevel value

commandBufferCount :: Word32

commandBufferCount is the number of command buffers to allocate from the pool.

commandBufferCount must be greater than 0

data CommandBufferInheritanceInfo (es :: [Type])

VkCommandBufferInheritanceInfo - Structure specifying command buffer inheritance info

Valid Usage

If the inherited queries feature is not enabled, occlusionQueryEnable must be FALSE

If the inherited queries feature is enabled, queryFlags must be a valid combination of QueryControlFlagBits values

If the inherited queries feature is not enabled, queryFlags must be 0

If the pipeline statistics queries feature is enabled, pipelineStatistics must be a valid combination of QueryPipelineStatisticFlagBits values

If the pipeline statistics queries feature is not enabled, pipelineStatistics must be 0

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_COMMAND_BUFFER_INHERITANCE_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of CommandBufferInheritanceConditionalRenderingInfoEXT or CommandBufferInheritanceRenderPassTransformInfoQCOM

The sType value of each struct in the pNext chain must be unique

Both of framebuffer, and renderPass that are valid handles of non-ignored parameters must have been created, allocated, or retrieved from the same Device

See Also

Bool32, CommandBufferBeginInfo, Framebuffer, QueryControlFlags, QueryPipelineStatisticFlags, RenderPass, StructureType

Constructors

CommandBufferInheritanceInfo

Fields

next :: Chain es

pNext is NULL or a pointer to a structure extending this structure.

renderPass :: RenderPass

renderPass is a RenderPass object defining which render passes the CommandBuffer will be compatible with and can be executed within. If the CommandBuffer will not be executed within a render pass instance, renderPass is ignored.

subpass :: Word32

subpass is the index of the subpass within the render pass instance that the CommandBuffer will be executed within. If the CommandBuffer will not be executed within a render pass instance, subpass is ignored.

framebuffer :: Framebuffer

framebuffer optionally refers to the Framebuffer object that the CommandBuffer will be rendering to if it is executed within a render pass instance. It can be NULL_HANDLE if the framebuffer is not known, or if the CommandBuffer will not be executed within a render pass instance.

Note

Specifying the exact framebuffer that the secondary command buffer will be executed with may result in better performance at command buffer execution time.

occlusionQueryEnable :: Bool

occlusionQueryEnable specifies whether the command buffer can be executed while an occlusion query is active in the primary command buffer. If this is TRUE, then this command buffer can be executed whether the primary command buffer has an occlusion query active or not. If this is FALSE, then the primary command buffer must not have an occlusion query active.

queryFlags :: QueryControlFlags

queryFlags specifies the query flags that can be used by an active occlusion query in the primary command buffer when this secondary command buffer is executed. If this value includes the QUERY_CONTROL_PRECISE_BIT bit, then the active query can return boolean results or actual sample counts. If this bit is not set, then the active query must not use the QUERY_CONTROL_PRECISE_BIT bit.

pipelineStatistics :: QueryPipelineStatisticFlags

pipelineStatistics is a bitmask of QueryPipelineStatisticFlagBits specifying the set of pipeline statistics that can be counted by an active query in the primary command buffer when this secondary command buffer is executed. If this value includes a given bit, then this command buffer can be executed whether the primary command buffer has a pipeline statistics query active that includes this bit or not. If this value excludes a given bit, then the active pipeline statistics query must not be from a query pool that counts that statistic.

data CommandBufferBeginInfo (es :: [Type])

VkCommandBufferBeginInfo - Structure specifying a command buffer begin operation

Valid Usage

If flags contains COMMAND_BUFFER_USAGE_RENDER_PASS_CONTINUE_BIT, the renderPass member of pInheritanceInfo must be a valid RenderPass

If flags contains COMMAND_BUFFER_USAGE_RENDER_PASS_CONTINUE_BIT, the subpass member of pInheritanceInfo must be a valid subpass index within the renderPass member of pInheritanceInfo

If flags contains COMMAND_BUFFER_USAGE_RENDER_PASS_CONTINUE_BIT, the framebuffer member of pInheritanceInfo must be either NULL_HANDLE, or a valid Framebuffer that is compatible with the renderPass member of pInheritanceInfo

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO

pNext must be NULL or a pointer to a valid instance of DeviceGroupCommandBufferBeginInfo

The sType value of each struct in the pNext chain must be unique

flags must be a valid combination of CommandBufferUsageFlagBits values

See Also

CommandBufferInheritanceInfo, CommandBufferUsageFlags, StructureType, beginCommandBuffer

Constructors

CommandBufferBeginInfo

Fields

next :: Chain es

pNext is NULL or a pointer to a structure extending this structure.

flags :: CommandBufferUsageFlags

flags is a bitmask of CommandBufferUsageFlagBits specifying usage behavior for the command buffer.

inheritanceInfo :: Maybe (SomeStruct CommandBufferInheritanceInfo)

pInheritanceInfo is a pointer to a CommandBufferInheritanceInfo structure, used if commandBuffer is a secondary command buffer. If this is a primary command buffer, then this value is ignored.

data CommandBuffer

VkCommandBuffer - Opaque handle to a command buffer object

See Also

SubmitInfo, allocateCommandBuffers, beginCommandBuffer, cmdBeginConditionalRenderingEXT, cmdBeginDebugUtilsLabelEXT, cmdBeginQuery, cmdBeginQueryIndexedEXT, cmdBeginRenderPass, cmdBeginRenderPass2, cmdBeginRenderPass2KHR, cmdBeginTransformFeedbackEXT, cmdBindDescriptorSets, cmdBindIndexBuffer, cmdBindPipeline, cmdBindPipelineShaderGroupNV, cmdBindShadingRateImageNV, cmdBindTransformFeedbackBuffersEXT, cmdBindVertexBuffers, cmdBindVertexBuffers2EXT, cmdBlitImage, cmdBuildAccelerationStructureIndirectKHR, cmdBuildAccelerationStructureKHR, cmdBuildAccelerationStructureNV, cmdClearAttachments, cmdClearColorImage, cmdClearDepthStencilImage, cmdCopyAccelerationStructureKHR, cmdCopyAccelerationStructureNV, cmdCopyAccelerationStructureToMemoryKHR, cmdCopyBuffer, cmdCopyBufferToImage, cmdCopyImage, cmdCopyImageToBuffer, cmdCopyMemoryToAccelerationStructureKHR, cmdCopyQueryPoolResults, cmdDebugMarkerBeginEXT, cmdDebugMarkerEndEXT, cmdDebugMarkerInsertEXT, cmdDispatch, cmdDispatchBase, cmdDispatchBaseKHR, cmdDispatchIndirect, cmdDraw, cmdDrawIndexed, cmdDrawIndexedIndirect, cmdDrawIndexedIndirectCount, cmdDrawIndexedIndirectCountAMD, cmdDrawIndexedIndirectCountKHR, cmdDrawIndirect, cmdDrawIndirectByteCountEXT, cmdDrawIndirectCount, cmdDrawIndirectCountAMD, cmdDrawIndirectCountKHR, cmdDrawMeshTasksIndirectCountNV, cmdDrawMeshTasksIndirectNV, cmdDrawMeshTasksNV, cmdEndConditionalRenderingEXT, cmdEndDebugUtilsLabelEXT, cmdEndQuery, cmdEndQueryIndexedEXT, cmdEndRenderPass, cmdEndRenderPass2, cmdEndRenderPass2KHR, cmdEndTransformFeedbackEXT, cmdExecuteCommands, cmdExecuteGeneratedCommandsNV, cmdFillBuffer, cmdInsertDebugUtilsLabelEXT, cmdNextSubpass, cmdNextSubpass2, cmdNextSubpass2KHR, cmdPipelineBarrier, cmdPreprocessGeneratedCommandsNV, cmdPushConstants, cmdPushDescriptorSetKHR, cmdPushDescriptorSetWithTemplateKHR, cmdResetEvent, cmdResetQueryPool, cmdResolveImage, cmdSetBlendConstants, cmdSetCheckpointNV, cmdSetCoarseSampleOrderNV, cmdSetCullModeEXT, cmdSetDepthBias, cmdSetDepthBounds, cmdSetDepthBoundsTestEnableEXT, cmdSetDepthCompareOpEXT, cmdSetDepthTestEnableEXT, cmdSetDepthWriteEnableEXT, cmdSetDeviceMask, cmdSetDeviceMaskKHR, cmdSetDiscardRectangleEXT, cmdSetEvent, cmdSetExclusiveScissorNV, cmdSetFrontFaceEXT, cmdSetLineStippleEXT, cmdSetLineWidth, cmdSetPerformanceMarkerINTEL, cmdSetPerformanceOverrideINTEL, cmdSetPerformanceStreamMarkerINTEL, cmdSetPrimitiveTopologyEXT, cmdSetSampleLocationsEXT, cmdSetScissor, cmdSetScissorWithCountEXT, cmdSetStencilCompareMask, cmdSetStencilOpEXT, cmdSetStencilReference, cmdSetStencilTestEnableEXT, cmdSetStencilWriteMask, cmdSetViewport, cmdSetViewportShadingRatePaletteNV, cmdSetViewportWScalingNV, cmdSetViewportWithCountEXT, cmdTraceRaysIndirectKHR, cmdTraceRaysKHR, cmdTraceRaysNV, cmdUpdateBuffer, cmdWaitEvents, cmdWriteAccelerationStructuresPropertiesKHR, cmdWriteAccelerationStructuresPropertiesNV, cmdWriteBufferMarkerAMD, cmdWriteTimestamp, endCommandBuffer, freeCommandBuffers, resetCommandBuffer

Constructors

CommandBuffer

Fields

commandBufferHandle :: Ptr CommandBuffer_T

deviceCmds :: DeviceCmds

newtype CommandBufferLevel

VkCommandBufferLevel - Enumerant specifying a command buffer level

See Also

CommandBufferAllocateInfo

Constructors

CommandBufferLevel Int32

Bundled Patterns

pattern COMMAND_BUFFER_LEVEL_PRIMARY :: CommandBufferLevel
COMMAND_BUFFER_LEVEL_PRIMARY specifies a primary command buffer.

pattern COMMAND_BUFFER_LEVEL_SECONDARY :: CommandBufferLevel
COMMAND_BUFFER_LEVEL_SECONDARY specifies a secondary command buffer.

newtype QueryControlFlagBits

VkQueryControlFlagBits - Bitmask specifying constraints on a query

See Also

QueryControlFlags

Constructors

QueryControlFlagBits Flags

Bundled Patterns

pattern QUERY_CONTROL_PRECISE_BIT :: QueryControlFlagBits
QUERY_CONTROL_PRECISE_BIT specifies the precision of occlusion queries.

type QueryControlFlags = QueryControlFlagBits

newtype CommandBufferUsageFlagBits

VkCommandBufferUsageFlagBits - Bitmask specifying usage behavior for command buffer

See Also

CommandBufferUsageFlags

Constructors

CommandBufferUsageFlagBits Flags

Bundled Patterns

pattern COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT :: CommandBufferUsageFlagBits
COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT specifies that each recording of the command buffer will only be submitted once, and the command buffer will be reset and recorded again between each submission.

pattern COMMAND_BUFFER_USAGE_RENDER_PASS_CONTINUE_BIT :: CommandBufferUsageFlagBits
COMMAND_BUFFER_USAGE_RENDER_PASS_CONTINUE_BIT specifies that a secondary command buffer is considered to be entirely inside a render pass. If this is a primary command buffer, then this bit is ignored.

pattern COMMAND_BUFFER_USAGE_SIMULTANEOUS_USE_BIT :: CommandBufferUsageFlagBits
COMMAND_BUFFER_USAGE_SIMULTANEOUS_USE_BIT specifies that a command buffer can be resubmitted to a queue while it is in the /pending state/, and recorded into multiple primary command buffers.

type CommandBufferUsageFlags = CommandBufferUsageFlagBits

newtype CommandBufferResetFlagBits

VkCommandBufferResetFlagBits - Bitmask controlling behavior of a command buffer reset

See Also

CommandBufferResetFlags

Constructors

CommandBufferResetFlagBits Flags

Bundled Patterns

pattern COMMAND_BUFFER_RESET_RELEASE_RESOURCES_BIT :: CommandBufferResetFlagBits
COMMAND_BUFFER_RESET_RELEASE_RESOURCES_BIT specifies that most or all memory resources currently owned by the command buffer should be returned to the parent command pool. If this flag is not set, then the command buffer may hold onto memory resources and reuse them when recording commands. commandBuffer is moved to the initial state.

type CommandBufferResetFlags = CommandBufferResetFlagBits