Vulkan.Core10.CommandPool

Documentation

createCommandPool

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that creates the command pool.

-> CommandPoolCreateInfo
pCreateInfo is a pointer to a CommandPoolCreateInfo structure specifying the state of the command pool object.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io CommandPool

vkCreateCommandPool - Create a new command pool object

Valid Usage

pCreateInfo->queueFamilyIndex must be the index of a queue family available in the logical device device

Valid Usage (Implicit)

device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid CommandPoolCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pCommandPool must be a valid pointer to a CommandPool handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, CommandPool, CommandPoolCreateInfo, Device

withCommandPool :: forall io r. MonadIO io => Device -> CommandPoolCreateInfo -> Maybe AllocationCallbacks -> (io CommandPool -> (CommandPool -> io ()) -> r) -> r

A convenience wrapper to make a compatible pair of calls to createCommandPool and destroyCommandPool

To ensure that destroyCommandPool is always called: pass bracket (or the allocate function from your favourite resource management library) as the first argument. To just extract the pair pass (,) as the first argument.

destroyCommandPool

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that destroys the command pool.

-> CommandPool
commandPool is the handle of the command pool to destroy.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io ()

vkDestroyCommandPool - Destroy a command pool object

Description

When a pool is destroyed, all command buffers allocated from the pool are freed.

Any primary command buffer allocated from another CommandPool that is in the recording or executable state and has a secondary command buffer allocated from commandPool recorded into it, becomes invalid.

Valid Usage

All CommandBuffer objects allocated from commandPool must not be in the pending state

If AllocationCallbacks were provided when commandPool was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when commandPool was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid Device handle

If commandPool is not NULL_HANDLE, commandPool must be a valid CommandPool handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If commandPool is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to commandPool must be externally synchronized

See Also

AllocationCallbacks, CommandPool, Device

resetCommandPool

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that owns the command pool.

-> CommandPool
commandPool is the command pool to reset.

-> CommandPoolResetFlags
flags is a bitmask of CommandPoolResetFlagBits controlling the reset operation.

-> io ()

vkResetCommandPool - Reset a command pool

Description

Resetting a command pool recycles all of the resources from all of the command buffers allocated from the command pool back to the command pool. All command buffers that have been allocated from the command pool are put in the initial state.

Any primary command buffer allocated from another CommandPool that is in the recording or executable state and has a secondary command buffer allocated from commandPool recorded into it, becomes invalid.

Valid Usage

All CommandBuffer objects allocated from commandPool must not be in the pending state

Valid Usage (Implicit)

device must be a valid Device handle

commandPool must be a valid CommandPool handle

flags must be a valid combination of CommandPoolResetFlagBits values

commandPool must have been created, allocated, or retrieved from device

Host Synchronization

Host access to commandPool must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

CommandPool, CommandPoolResetFlags, Device

data CommandPoolCreateInfo

VkCommandPoolCreateInfo - Structure specifying parameters of a newly created command pool

Valid Usage

If the protected memory feature is not enabled, the COMMAND_POOL_CREATE_PROTECTED_BIT bit of flags must not be set

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO

pNext must be NULL

flags must be a valid combination of CommandPoolCreateFlagBits values

See Also

CommandPoolCreateFlags, StructureType, createCommandPool

Constructors

CommandPoolCreateInfo

Fields

flags :: CommandPoolCreateFlags

flags is a bitmask of CommandPoolCreateFlagBits indicating usage behavior for the pool and command buffers allocated from it.

queueFamilyIndex :: Word32

queueFamilyIndex designates a queue family as described in section Queue Family Properties. All command buffers allocated from this command pool must be submitted on queues from the same queue family.

newtype CommandPool

VkCommandPool - Opaque handle to a command pool object

See Also

CommandBufferAllocateInfo, createCommandPool, destroyCommandPool, freeCommandBuffers, resetCommandPool, trimCommandPool, trimCommandPoolKHR

Constructors

CommandPool Word64

newtype CommandPoolCreateFlagBits

VkCommandPoolCreateFlagBits - Bitmask specifying usage behavior for a command pool

See Also

CommandPoolCreateFlags

Constructors

CommandPoolCreateFlagBits Flags

Bundled Patterns

pattern COMMAND_POOL_CREATE_TRANSIENT_BIT :: CommandPoolCreateFlagBits
COMMAND_POOL_CREATE_TRANSIENT_BIT specifies that command buffers allocated from the pool will be short-lived, meaning that they will be reset or freed in a relatively short timeframe. This flag may be used by the implementation to control memory allocation behavior within the pool.

pattern COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT :: CommandPoolCreateFlagBits
COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT allows any command buffer allocated from a pool to be individually reset to the initial state; either by calling resetCommandBuffer, or via the implicit reset when calling beginCommandBuffer. If this flag is not set on a pool, then resetCommandBuffer must not be called for any command buffer allocated from that pool.

pattern COMMAND_POOL_CREATE_PROTECTED_BIT :: CommandPoolCreateFlagBits
COMMAND_POOL_CREATE_PROTECTED_BIT specifies that command buffers allocated from the pool are protected command buffers.

type CommandPoolCreateFlags = CommandPoolCreateFlagBits

newtype CommandPoolResetFlagBits

VkCommandPoolResetFlagBits - Bitmask controlling behavior of a command pool reset

See Also

CommandPoolResetFlags

Constructors

CommandPoolResetFlagBits Flags

Bundled Patterns

pattern COMMAND_POOL_RESET_RELEASE_RESOURCES_BIT :: CommandPoolResetFlagBits
COMMAND_POOL_RESET_RELEASE_RESOURCES_BIT specifies that resetting a command pool recycles all of the resources from the command pool back to the system.

type CommandPoolResetFlags = CommandPoolResetFlagBits