Vulkan.Core10.DescriptorSet

Documentation

createDescriptorSetLayout

Arguments

:: forall a io. (Extendss DescriptorSetLayoutCreateInfo a, PokeChain a, MonadIO io)

=> Device
device is the logical device that creates the descriptor set layout.

-> DescriptorSetLayoutCreateInfo a
pCreateInfo is a pointer to a DescriptorSetLayoutCreateInfo structure specifying the state of the descriptor set layout object.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io DescriptorSetLayout

vkCreateDescriptorSetLayout - Create a new descriptor set layout

Valid Usage (Implicit)

device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid DescriptorSetLayoutCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pSetLayout must be a valid pointer to a DescriptorSetLayout handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, DescriptorSetLayout, DescriptorSetLayoutCreateInfo, Device

withDescriptorSetLayout :: forall a io r. (Extendss DescriptorSetLayoutCreateInfo a, PokeChain a, MonadIO io) => Device -> DescriptorSetLayoutCreateInfo a -> Maybe AllocationCallbacks -> (io DescriptorSetLayout -> (DescriptorSetLayout -> io ()) -> r) -> r

A convenience wrapper to make a compatible pair of calls to createDescriptorSetLayout and destroyDescriptorSetLayout

To ensure that destroyDescriptorSetLayout is always called: pass bracket (or the allocate function from your favourite resource management library) as the first argument. To just extract the pair pass (,) as the first argument.

destroyDescriptorSetLayout

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that destroys the descriptor set layout.

-> DescriptorSetLayout
descriptorSetLayout is the descriptor set layout to destroy.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io ()

vkDestroyDescriptorSetLayout - Destroy a descriptor set layout object

Valid Usage

If AllocationCallbacks were provided when descriptorSetLayout was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when descriptorSetLayout was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid Device handle

If descriptorSetLayout is not NULL_HANDLE, descriptorSetLayout must be a valid DescriptorSetLayout handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If descriptorSetLayout is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to descriptorSetLayout must be externally synchronized

See Also

AllocationCallbacks, DescriptorSetLayout, Device

createDescriptorPool

Arguments

:: forall a io. (Extendss DescriptorPoolCreateInfo a, PokeChain a, MonadIO io)

=> Device
device is the logical device that creates the descriptor pool.

-> DescriptorPoolCreateInfo a
pCreateInfo is a pointer to a DescriptorPoolCreateInfo structure specifying the state of the descriptor pool object.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io DescriptorPool

vkCreateDescriptorPool - Creates a descriptor pool object

Description

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

The created descriptor pool is returned in pDescriptorPool.

Valid Usage (Implicit)

device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid DescriptorPoolCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pDescriptorPool must be a valid pointer to a DescriptorPool handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_FRAGMENTATION_EXT

See Also

AllocationCallbacks, DescriptorPool, DescriptorPoolCreateInfo, Device

withDescriptorPool :: forall a io r. (Extendss DescriptorPoolCreateInfo a, PokeChain a, MonadIO io) => Device -> DescriptorPoolCreateInfo a -> Maybe AllocationCallbacks -> (io DescriptorPool -> (DescriptorPool -> io ()) -> r) -> r

A convenience wrapper to make a compatible pair of calls to createDescriptorPool and destroyDescriptorPool

To ensure that destroyDescriptorPool is always called: pass bracket (or the allocate function from your favourite resource management library) as the first argument. To just extract the pair pass (,) as the first argument.

destroyDescriptorPool

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that destroys the descriptor pool.

-> DescriptorPool
descriptorPool is the descriptor pool to destroy.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io ()

vkDestroyDescriptorPool - Destroy a descriptor pool object

Description

When a pool is destroyed, all descriptor sets allocated from the pool are implicitly freed and become invalid. Descriptor sets allocated from a given pool do not need to be freed before destroying that descriptor pool.

Valid Usage

All submitted commands that refer to descriptorPool (via any allocated descriptor sets) must have completed execution

If AllocationCallbacks were provided when descriptorPool was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when descriptorPool was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid Device handle

If descriptorPool is not NULL_HANDLE, descriptorPool must be a valid DescriptorPool handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If descriptorPool is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to descriptorPool must be externally synchronized

See Also

AllocationCallbacks, DescriptorPool, Device

resetDescriptorPool

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that owns the descriptor pool.

-> DescriptorPool
descriptorPool is the descriptor pool to be reset.

-> DescriptorPoolResetFlags
flags is reserved for future use.

-> io ()

vkResetDescriptorPool - Resets a descriptor pool object

Description

Resetting a descriptor pool recycles all of the resources from all of the descriptor sets allocated from the descriptor pool back to the descriptor pool, and the descriptor sets are implicitly freed.

Valid Usage

All uses of descriptorPool (via any allocated descriptor sets) must have completed execution

Valid Usage (Implicit)

device must be a valid Device handle

descriptorPool must be a valid DescriptorPool handle

flags must be 0

descriptorPool must have been created, allocated, or retrieved from device

Host Synchronization

Host access to descriptorPool must be externally synchronized

Host access to any DescriptorSet objects allocated from descriptorPool must be externally synchronized

Return Codes

Success

SUCCESS

See Also

DescriptorPool, DescriptorPoolResetFlags, Device

allocateDescriptorSets

Arguments

:: forall a io. (Extendss DescriptorSetAllocateInfo a, PokeChain a, MonadIO io)

=> Device
device is the logical device that owns the descriptor pool.

-> DescriptorSetAllocateInfo a
pAllocateInfo is a pointer to a DescriptorSetAllocateInfo structure describing parameters of the allocation.

-> io ("descriptorSets" ::: Vector DescriptorSet)

vkAllocateDescriptorSets - Allocate one or more descriptor sets

Description

The allocated descriptor sets are returned in pDescriptorSets.

When a descriptor set is allocated, the initial state is largely uninitialized and all descriptors are undefined. Descriptors also become undefined if the underlying resource is destroyed. Descriptor sets containing undefined descriptors can still be bound and used, subject to the following conditions:

For descriptor set bindings created with the DESCRIPTOR_BINDING_PARTIALLY_BOUND_BIT bit set, all descriptors in that binding that are dynamically used must have been populated before the descriptor set is consumed.

For descriptor set bindings created without the DESCRIPTOR_BINDING_PARTIALLY_BOUND_BIT bit set, all descriptors in that binding that are statically used must have been populated before the descriptor set is consumed.

Descriptor bindings with descriptor type of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT can be undefined when the descriptor set is consumed; though values in that block will be undefined.

Entries that are not used by a pipeline can have undefined descriptors.

If a call to allocateDescriptorSets would cause the total number of descriptor sets allocated from the pool to exceed the value of DescriptorPoolCreateInfo::maxSets used to create pAllocateInfo->descriptorPool, then the allocation may fail due to lack of space in the descriptor pool. Similarly, the allocation may fail due to lack of space if the call to allocateDescriptorSets would cause the number of any given descriptor type to exceed the sum of all the descriptorCount members of each element of DescriptorPoolCreateInfo::pPoolSizes with a member equal to that type.

Additionally, the allocation may also fail if a call to allocateDescriptorSets would cause the total number of inline uniform block bindings allocated from the pool to exceed the value of DescriptorPoolInlineUniformBlockCreateInfoEXT::maxInlineUniformBlockBindings used to create the descriptor pool.

If the allocation fails due to no more space in the descriptor pool, and not because of system or device memory exhaustion, then ERROR_OUT_OF_POOL_MEMORY must be returned.

allocateDescriptorSets can be used to create multiple descriptor sets. If the creation of any of those descriptor sets fails, then the implementation must destroy all successfully created descriptor set objects from this command, set all entries of the pDescriptorSets array to NULL_HANDLE and return the error.

Valid Usage (Implicit)

device must be a valid Device handle

pAllocateInfo must be a valid pointer to a valid DescriptorSetAllocateInfo structure

pDescriptorSets must be a valid pointer to an array of pAllocateInfo->descriptorSetCount DescriptorSet handles

The value referenced by pAllocateInfo->descriptorSetCount must be greater than 0

Host Synchronization

Host access to pAllocateInfo->descriptorPool must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_FRAGMENTED_POOL

ERROR_OUT_OF_POOL_MEMORY

See Also

DescriptorSet, DescriptorSetAllocateInfo, Device

withDescriptorSets :: forall a io r. (Extendss DescriptorSetAllocateInfo a, PokeChain a, MonadIO io) => Device -> DescriptorSetAllocateInfo a -> (io (Vector DescriptorSet) -> (Vector DescriptorSet -> io ()) -> r) -> r

A convenience wrapper to make a compatible pair of calls to allocateDescriptorSets and freeDescriptorSets

To ensure that freeDescriptorSets is always called: pass bracket (or the allocate function from your favourite resource management library) as the first argument. To just extract the pair pass (,) as the first argument.

freeDescriptorSets

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that owns the descriptor pool.

-> DescriptorPool
descriptorPool is the descriptor pool from which the descriptor sets were allocated.

-> ("descriptorSets" ::: Vector DescriptorSet)
pDescriptorSets is a pointer to an array of handles to DescriptorSet objects.

-> io ()

vkFreeDescriptorSets - Free one or more descriptor sets

Description

After calling freeDescriptorSets, all descriptor sets in pDescriptorSets are invalid.

Valid Usage

All submitted commands that refer to any element of pDescriptorSets must have completed execution

pDescriptorSets must be a valid pointer to an array of descriptorSetCount DescriptorSet handles, each element of which must either be a valid handle or NULL_HANDLE

Each valid handle in pDescriptorSets must have been allocated from descriptorPool

descriptorPool must have been created with the DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT flag

Valid Usage (Implicit)

device must be a valid Device handle

descriptorPool must be a valid DescriptorPool handle

descriptorSetCount must be greater than 0

descriptorPool must have been created, allocated, or retrieved from device

Each element of pDescriptorSets that is a valid handle must have been created, allocated, or retrieved from descriptorPool

Host Synchronization

Host access to descriptorPool must be externally synchronized

Host access to each member of pDescriptorSets must be externally synchronized

Return Codes

Success

SUCCESS

See Also

DescriptorPool, DescriptorSet, Device

updateDescriptorSets

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that updates the descriptor sets.

-> ("descriptorWrites" ::: Vector (SomeStruct WriteDescriptorSet))
pDescriptorWrites is a pointer to an array of WriteDescriptorSet structures describing the descriptor sets to write to.

-> ("descriptorCopies" ::: Vector CopyDescriptorSet)
pDescriptorCopies is a pointer to an array of CopyDescriptorSet structures describing the descriptor sets to copy between.

-> io ()

vkUpdateDescriptorSets - Update the contents of a descriptor set object

Description

The operations described by pDescriptorWrites are performed first, followed by the operations described by pDescriptorCopies. Within each array, the operations are performed in the order they appear in the array.

Each element in the pDescriptorWrites array describes an operation updating the descriptor set using descriptors for resources specified in the structure.

Each element in the pDescriptorCopies array is a CopyDescriptorSet structure describing an operation copying descriptors between sets.

If the dstSet member of any element of pDescriptorWrites or pDescriptorCopies is bound, accessed, or modified by any command that was recorded to a command buffer which is currently in the recording or executable state, and any of the descriptor bindings that are updated were not created with the DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT or DESCRIPTOR_BINDING_UPDATE_UNUSED_WHILE_PENDING_BIT bits set, that command buffer becomes invalid.

Valid Usage

Descriptor bindings updated by this command which were created without the DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT or DESCRIPTOR_BINDING_UPDATE_UNUSED_WHILE_PENDING_BIT bits set must not be used by any command that was recorded to a command buffer which is in the pending state

Valid Usage (Implicit)

device must be a valid Device handle

If descriptorWriteCount is not 0, pDescriptorWrites must be a valid pointer to an array of descriptorWriteCount valid WriteDescriptorSet structures

If descriptorCopyCount is not 0, pDescriptorCopies must be a valid pointer to an array of descriptorCopyCount valid CopyDescriptorSet structures

Host Synchronization

Host access to pDescriptorWrites[].dstSet must be externally synchronized

Host access to pDescriptorCopies[].dstSet must be externally synchronized

See Also

CopyDescriptorSet, Device, WriteDescriptorSet

data DescriptorBufferInfo

VkDescriptorBufferInfo - Structure specifying descriptor buffer info

Description

Note

When setting range to WHOLE_SIZE, the effective range must not be larger than the maximum range for the descriptor type (maxUniformBufferRange or maxStorageBufferRange). This means that WHOLE_SIZE is not typically useful in the common case where uniform buffer descriptors are suballocated from a buffer that is much larger than maxUniformBufferRange.

For DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC and DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC descriptor types, offset is the base offset from which the dynamic offset is applied and range is the static size used for all dynamic offsets.

Valid Usage

offset must be less than the size of buffer

If range is not equal to WHOLE_SIZE, range must be greater than 0

If range is not equal to WHOLE_SIZE, range must be less than or equal to the size of buffer minus offset

If the nullDescriptor feature is not enabled, buffer must not be NULL_HANDLE

If buffer is NULL_HANDLE, offset must be zero and range must be WHOLE_SIZE

Valid Usage (Implicit)

If buffer is not NULL_HANDLE, buffer must be a valid Buffer handle

See Also

Buffer, DeviceSize, WriteDescriptorSet

Constructors

DescriptorBufferInfo

Fields

buffer :: Buffer

buffer is NULL_HANDLE or the buffer resource.

offset :: DeviceSize

offset is the offset in bytes from the start of buffer. Access to buffer memory via this descriptor uses addressing that is relative to this starting offset.

range :: DeviceSize

range is the size in bytes that is used for this descriptor update, or WHOLE_SIZE to use the range from offset to the end of the buffer.

data DescriptorImageInfo

VkDescriptorImageInfo - Structure specifying descriptor image info

Description

Members of DescriptorImageInfo that are not used in an update (as described above) are ignored.

Valid Usage

imageView must not be 2D or 2D array image view created from a 3D image

If imageView is created from a depth/stencil image, the aspectMask used to create the imageView must include either IMAGE_ASPECT_DEPTH_BIT or IMAGE_ASPECT_STENCIL_BIT but not both

imageLayout must match the actual ImageLayout of each subresource accessible from imageView at the time this descriptor is accessed as defined by the image layout matching rules

If sampler is used and the Format of the image is a multi-planar format, the image must have been created with IMAGE_CREATE_MUTABLE_FORMAT_BIT, and the aspectMask of the imageView must be IMAGE_ASPECT_PLANE_0_BIT, IMAGE_ASPECT_PLANE_1_BIT or (for three-plane formats only) IMAGE_ASPECT_PLANE_2_BIT

Valid Usage (Implicit)

Both of imageView, and sampler that are valid handles of non-ignored parameters must have been created, allocated, or retrieved from the same Device

See Also

ImageLayout, ImageView, Sampler, WriteDescriptorSet

Constructors

DescriptorImageInfo

Fields

sampler :: Sampler

sampler is a sampler handle, and is used in descriptor updates for types DESCRIPTOR_TYPE_SAMPLER and DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER if the binding being updated does not use immutable samplers.

imageView :: ImageView

imageView is NULL_HANDLE or an image view handle, and is used in descriptor updates for types DESCRIPTOR_TYPE_SAMPLED_IMAGE, DESCRIPTOR_TYPE_STORAGE_IMAGE, DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, and DESCRIPTOR_TYPE_INPUT_ATTACHMENT.

imageLayout :: ImageLayout

imageLayout is the layout that the image subresources accessible from imageView will be in at the time this descriptor is accessed. imageLayout is used in descriptor updates for types DESCRIPTOR_TYPE_SAMPLED_IMAGE, DESCRIPTOR_TYPE_STORAGE_IMAGE, DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, and DESCRIPTOR_TYPE_INPUT_ATTACHMENT.

data WriteDescriptorSet (es :: [Type])

VkWriteDescriptorSet - Structure specifying the parameters of a descriptor set write operation

Description

Only one of pImageInfo, pBufferInfo, or pTexelBufferView members is used according to the descriptor type specified in the descriptorType member of the containing WriteDescriptorSet structure, or none of them in case descriptorType is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT, in which case the source data for the descriptor writes is taken from the WriteDescriptorSetInlineUniformBlockEXT structure included in the pNext chain of WriteDescriptorSet, or if descriptorType is DESCRIPTOR_TYPE_ACCELERATION_STRUCTURE_KHR, in which case the source data for the descriptor writes is taken from the WriteDescriptorSetAccelerationStructureKHR structure in the pNext chain of WriteDescriptorSet, as specified below.

If the nullDescriptor feature is enabled, the buffer, imageView, or bufferView can be NULL_HANDLE. Loads from a null descriptor return zero values and stores and atomics to a null descriptor are discarded.

If the dstBinding has fewer than descriptorCount array elements remaining starting from dstArrayElement, then the remainder will be used to update the subsequent binding - dstBinding+1 starting at array element zero. If a binding has a descriptorCount of zero, it is skipped. This behavior applies recursively, with the update affecting consecutive bindings as needed to update all descriptorCount descriptors.

Note

The same behavior applies to bindings with a descriptor type of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT where descriptorCount specifies the number of bytes to update while dstArrayElement specifies the starting byte offset, thus in this case if the dstBinding has a smaller byte size than the sum of dstArrayElement and descriptorCount, then the remainder will be used to update the subsequent binding - dstBinding+1 starting at offset zero. This falls out as a special case of the above rule.

Valid Usage

dstBinding must be less than or equal to the maximum value of binding of all DescriptorSetLayoutBinding structures specified when dstSet’s descriptor set layout was created

dstBinding must be a binding with a non-zero descriptorCount

All consecutive bindings updated via a single WriteDescriptorSet structure, except those with a descriptorCount of zero, must have identical descriptorType and stageFlags

All consecutive bindings updated via a single WriteDescriptorSet structure, except those with a descriptorCount of zero, must all either use immutable samplers or must all not use immutable samplers

descriptorType must match the type of dstBinding within dstSet

dstSet must be a valid DescriptorSet handle

The sum of dstArrayElement and descriptorCount must be less than or equal to the number of array elements in the descriptor set binding specified by dstBinding, and all applicable consecutive bindings, as described by https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#descriptorsets-updates-consecutive

If descriptorType is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT, dstArrayElement must be an integer multiple of 4

If descriptorType is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT, descriptorCount must be an integer multiple of 4

If descriptorType is DESCRIPTOR_TYPE_SAMPLER, DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, DESCRIPTOR_TYPE_SAMPLED_IMAGE, DESCRIPTOR_TYPE_STORAGE_IMAGE, or DESCRIPTOR_TYPE_INPUT_ATTACHMENT, pImageInfo must be a valid pointer to an array of descriptorCount valid DescriptorImageInfo structures

If descriptorType is DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER or DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER, each element of pTexelBufferView must be either a valid BufferView handle or NULL_HANDLE

If descriptorType is DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER or DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER and the nullDescriptor feature is not enabled, each element of pTexelBufferView must not be NULL_HANDLE

If descriptorType is DESCRIPTOR_TYPE_UNIFORM_BUFFER, DESCRIPTOR_TYPE_STORAGE_BUFFER, DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC, pBufferInfo must be a valid pointer to an array of descriptorCount valid DescriptorBufferInfo structures

If descriptorType is DESCRIPTOR_TYPE_SAMPLER or DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, and dstSet was not allocated with a layout that included immutable samplers for dstBinding with descriptorType, the sampler member of each element of pImageInfo must be a valid Sampler object

If descriptorType is DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, DESCRIPTOR_TYPE_SAMPLED_IMAGE, DESCRIPTOR_TYPE_STORAGE_IMAGE, or DESCRIPTOR_TYPE_INPUT_ATTACHMENT, the imageView member of each element of pImageInfo must be either a valid ImageView handle or NULL_HANDLE

If descriptorType is DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, DESCRIPTOR_TYPE_SAMPLED_IMAGE, DESCRIPTOR_TYPE_STORAGE_IMAGE, or DESCRIPTOR_TYPE_INPUT_ATTACHMENT and the nullDescriptor feature is not enabled, the imageView member of each element of pImageInfo must not be NULL_HANDLE

If descriptorType is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT, the pNext chain must include a WriteDescriptorSetInlineUniformBlockEXT structure whose dataSize member equals descriptorCount

If descriptorType is DESCRIPTOR_TYPE_ACCELERATION_STRUCTURE_KHR, the pNext chain must include a WriteDescriptorSetAccelerationStructureKHR structure whose accelerationStructureCount member equals descriptorCount

If descriptorType is DESCRIPTOR_TYPE_SAMPLED_IMAGE, then the imageView member of each pImageInfo element must have been created without a SamplerYcbcrConversionInfo structure in its pNext chain

If descriptorType is DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, and if any element of pImageInfo has a imageView member that was created with a SamplerYcbcrConversionInfo structure in its pNext chain, then dstSet must have been allocated with a layout that included immutable samplers for dstBinding, and the corresponding immutable sampler must have been created with an identically defined SamplerYcbcrConversionInfo object

If descriptorType is DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, and dstSet was allocated with a layout that included immutable samplers for dstBinding, then the imageView member of each element of pImageInfo which corresponds to an immutable sampler that enables sampler Y′CBCR conversion must have been created with a SamplerYcbcrConversionInfo structure in its pNext chain with an identically defined SamplerYcbcrConversionInfo to the corresponding immutable sampler

If descriptorType is DESCRIPTOR_TYPE_STORAGE_IMAGE, for each descriptor that will be accessed via load or store operations the imageLayout member for corresponding elements of pImageInfo must be IMAGE_LAYOUT_GENERAL

If descriptorType is DESCRIPTOR_TYPE_UNIFORM_BUFFER or DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, the offset member of each element of pBufferInfo must be a multiple of PhysicalDeviceLimits::minUniformBufferOffsetAlignment

If descriptorType is DESCRIPTOR_TYPE_STORAGE_BUFFER or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC, the offset member of each element of pBufferInfo must be a multiple of PhysicalDeviceLimits::minStorageBufferOffsetAlignment

If descriptorType is DESCRIPTOR_TYPE_UNIFORM_BUFFER, DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, DESCRIPTOR_TYPE_STORAGE_BUFFER, or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC, and the buffer member of any element of pBufferInfo is the handle of a non-sparse buffer, then that buffer must be bound completely and contiguously to a single DeviceMemory object

If descriptorType is DESCRIPTOR_TYPE_UNIFORM_BUFFER or DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, the buffer member of each element of pBufferInfo must have been created with BUFFER_USAGE_UNIFORM_BUFFER_BIT set

If descriptorType is DESCRIPTOR_TYPE_STORAGE_BUFFER or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC, the buffer member of each element of pBufferInfo must have been created with BUFFER_USAGE_STORAGE_BUFFER_BIT set

If descriptorType is DESCRIPTOR_TYPE_UNIFORM_BUFFER or DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, the range member of each element of pBufferInfo, or the effective range if range is WHOLE_SIZE, must be less than or equal to PhysicalDeviceLimits::maxUniformBufferRange

If descriptorType is DESCRIPTOR_TYPE_STORAGE_BUFFER or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC, the range member of each element of pBufferInfo, or the effective range if range is WHOLE_SIZE, must be less than or equal to PhysicalDeviceLimits::maxStorageBufferRange

If descriptorType is DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER, the Buffer that each element of pTexelBufferView was created from must have been created with BUFFER_USAGE_UNIFORM_TEXEL_BUFFER_BIT set

If descriptorType is DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER, the Buffer that each element of pTexelBufferView was created from must have been created with BUFFER_USAGE_STORAGE_TEXEL_BUFFER_BIT set

If descriptorType is DESCRIPTOR_TYPE_STORAGE_IMAGE or DESCRIPTOR_TYPE_INPUT_ATTACHMENT, the imageView member of each element of pImageInfo must have been created with the identity swizzle

If descriptorType is DESCRIPTOR_TYPE_SAMPLED_IMAGE or DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, the imageView member of each element of pImageInfo must have been created with IMAGE_USAGE_SAMPLED_BIT set

If descriptorType is DESCRIPTOR_TYPE_SAMPLED_IMAGE or DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, the imageLayout member of each element of pImageInfo must be a member of the list given in Sampled Image or Combined Image Sampler, corresponding to its type

If descriptorType is DESCRIPTOR_TYPE_INPUT_ATTACHMENT, the imageView member of each element of pImageInfo must have been created with IMAGE_USAGE_INPUT_ATTACHMENT_BIT set

If descriptorType is DESCRIPTOR_TYPE_STORAGE_IMAGE, the imageView member of each element of pImageInfo must have been created with IMAGE_USAGE_STORAGE_BIT set

All consecutive bindings updated via a single WriteDescriptorSet structure, except those with a descriptorCount of zero, must have identical DescriptorBindingFlagBits

If descriptorType is DESCRIPTOR_TYPE_SAMPLER, then dstSet must not have been allocated with a layout that included immutable samplers for dstBinding

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of WriteDescriptorSetAccelerationStructureKHR or WriteDescriptorSetInlineUniformBlockEXT

The sType value of each struct in the pNext chain must be unique

descriptorType must be a valid DescriptorType value

descriptorCount must be greater than 0

Both of dstSet, and the elements of pTexelBufferView that are valid handles of non-ignored parameters must have been created, allocated, or retrieved from the same Device

See Also

BufferView, DescriptorBufferInfo, DescriptorImageInfo, DescriptorSet, DescriptorType, StructureType, cmdPushDescriptorSetKHR, updateDescriptorSets

Constructors

WriteDescriptorSet

Fields

next :: Chain es

pNext is NULL or a pointer to a structure extending this structure.

dstSet :: DescriptorSet

dstSet is the destination descriptor set to update.

dstBinding :: Word32

dstBinding is the descriptor binding within that set.

dstArrayElement :: Word32

dstArrayElement is the starting element in that array. If the descriptor binding identified by dstSet and dstBinding has a descriptor type of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then dstArrayElement specifies the starting byte offset within the binding.

descriptorCount :: Word32

descriptorCount is the number of descriptors to update (the number of elements in pImageInfo, pBufferInfo, or pTexelBufferView , or a value matching the dataSize member of a WriteDescriptorSetInlineUniformBlockEXT structure in the pNext chain , or a value matching the accelerationStructureCount of a WriteDescriptorSetAccelerationStructureKHR structure in the pNext chain ). If the descriptor binding identified by dstSet and dstBinding has a descriptor type of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then descriptorCount specifies the number of bytes to update.

descriptorType :: DescriptorType

descriptorType is a DescriptorType specifying the type of each descriptor in pImageInfo, pBufferInfo, or pTexelBufferView, as described below. It must be the same type as that specified in DescriptorSetLayoutBinding for dstSet at dstBinding. The type of the descriptor also controls which array the descriptors are taken from.

imageInfo :: Vector DescriptorImageInfo

pImageInfo is a pointer to an array of DescriptorImageInfo structures or is ignored, as described below.

bufferInfo :: Vector DescriptorBufferInfo

pBufferInfo is a pointer to an array of DescriptorBufferInfo structures or is ignored, as described below.

texelBufferView :: Vector BufferView

pTexelBufferView is a pointer to an array of BufferView handles as described in the Buffer Views section or is ignored, as described below.

data CopyDescriptorSet

VkCopyDescriptorSet - Structure specifying a copy descriptor set operation

Valid Usage

srcBinding must be a valid binding within srcSet

The sum of srcArrayElement and descriptorCount must be less than or equal to the number of array elements in the descriptor set binding specified by srcBinding, and all applicable consecutive bindings, as described by https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#descriptorsets-updates-consecutive

dstBinding must be a valid binding within dstSet

The sum of dstArrayElement and descriptorCount must be less than or equal to the number of array elements in the descriptor set binding specified by dstBinding, and all applicable consecutive bindings, as described by https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#descriptorsets-updates-consecutive

The type of dstBinding within dstSet must be equal to the type of srcBinding within srcSet

If srcSet is equal to dstSet, then the source and destination ranges of descriptors must not overlap, where the ranges may include array elements from consecutive bindings as described by https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#descriptorsets-updates-consecutive

If the descriptor type of the descriptor set binding specified by srcBinding is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT, srcArrayElement must be an integer multiple of 4

If the descriptor type of the descriptor set binding specified by dstBinding is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT, dstArrayElement must be an integer multiple of 4

If the descriptor type of the descriptor set binding specified by either srcBinding or dstBinding is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT, descriptorCount must be an integer multiple of 4

If srcSet’s layout was created with the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT flag set, then dstSet’s layout must also have been created with the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT flag set

If srcSet’s layout was created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT flag set, then dstSet’s layout must also have been created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT flag set

If the descriptor pool from which srcSet was allocated was created with the DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT flag set, then the descriptor pool from which dstSet was allocated must also have been created with the DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT flag set

If the descriptor pool from which srcSet was allocated was created without the DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT flag set, then the descriptor pool from which dstSet was allocated must also have been created without the DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT flag set

If the descriptor type of the descriptor set binding specified by dstBinding is DESCRIPTOR_TYPE_SAMPLER, then dstSet must not have been allocated with a layout that included immutable samplers for dstBinding

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_COPY_DESCRIPTOR_SET

pNext must be NULL

srcSet must be a valid DescriptorSet handle

dstSet must be a valid DescriptorSet handle

Both of dstSet, and srcSet must have been created, allocated, or retrieved from the same Device

See Also

DescriptorSet, StructureType, updateDescriptorSets

Constructors

CopyDescriptorSet

Fields

srcSet :: DescriptorSet

srcSet, srcBinding, and srcArrayElement are the source set, binding, and array element, respectively. If the descriptor binding identified by srcSet and srcBinding has a descriptor type of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then srcArrayElement specifies the starting byte offset within the binding to copy from.

srcBinding :: Word32

srcArrayElement :: Word32

dstSet :: DescriptorSet

dstSet, dstBinding, and dstArrayElement are the destination set, binding, and array element, respectively. If the descriptor binding identified by dstSet and dstBinding has a descriptor type of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then dstArrayElement specifies the starting byte offset within the binding to copy to.

dstBinding :: Word32

dstArrayElement :: Word32

descriptorCount :: Word32

descriptorCount is the number of descriptors to copy from the source to destination. If descriptorCount is greater than the number of remaining array elements in the source or destination binding, those affect consecutive bindings in a manner similar to WriteDescriptorSet above. If the descriptor binding identified by srcSet and srcBinding has a descriptor type of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then descriptorCount specifies the number of bytes to copy and the remaining array elements in the source or destination binding refer to the remaining number of bytes in those.

data DescriptorSetLayoutBinding

VkDescriptorSetLayoutBinding - Structure specifying a descriptor set layout binding

Description

pImmutableSamplers affects initialization of samplers. If descriptorType specifies a DESCRIPTOR_TYPE_SAMPLER or DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER type descriptor, then pImmutableSamplers can be used to initialize a set of immutable samplers. Immutable samplers are permanently bound into the set layout and must not be changed; updating a DESCRIPTOR_TYPE_SAMPLER descriptor with immutable samplers is not allowed and updates to a DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER descriptor with immutable samplers does not modify the samplers (the image views are updated, but the sampler updates are ignored). If pImmutableSamplers is not NULL, then it points to an array of sampler handles that will be copied into the set layout and used for the corresponding binding. Only the sampler handles are copied; the sampler objects must not be destroyed before the final use of the set layout and any descriptor pools and sets created using it. If pImmutableSamplers is NULL, then the sampler slots are dynamic and sampler handles must be bound into descriptor sets using this layout. If descriptorType is not one of these descriptor types, then pImmutableSamplers is ignored.

The above layout definition allows the descriptor bindings to be specified sparsely such that not all binding numbers between 0 and the maximum binding number need to be specified in the pBindings array. Bindings that are not specified have a descriptorCount and stageFlags of zero, and the value of descriptorType is undefined. However, all binding numbers between 0 and the maximum binding number in the DescriptorSetLayoutCreateInfo::pBindings array may consume memory in the descriptor set layout even if not all descriptor bindings are used, though it should not consume additional memory from the descriptor pool.

Note

The maximum binding number specified should be as compact as possible to avoid wasted memory.

Valid Usage

If descriptorType is DESCRIPTOR_TYPE_SAMPLER or DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, and descriptorCount is not 0 and pImmutableSamplers is not NULL, pImmutableSamplers must be a valid pointer to an array of descriptorCount valid Sampler handles

If descriptorType is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then descriptorCount must be a multiple of 4

If descriptorType is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then descriptorCount must be less than or equal to PhysicalDeviceInlineUniformBlockPropertiesEXT::maxInlineUniformBlockSize

If descriptorCount is not 0, stageFlags must be a valid combination of ShaderStageFlagBits values

If descriptorType is DESCRIPTOR_TYPE_INPUT_ATTACHMENT and descriptorCount is not 0, then stageFlags must be 0 or SHADER_STAGE_FRAGMENT_BIT

The sampler objects indicated by pImmutableSamplers must not have a borderColor with one of the values BORDER_COLOR_FLOAT_CUSTOM_EXT or BORDER_COLOR_INT_CUSTOM_EXT

Valid Usage (Implicit)

descriptorType must be a valid DescriptorType value

See Also

DescriptorSetLayoutCreateInfo, DescriptorType, Sampler, ShaderStageFlags

Constructors

DescriptorSetLayoutBinding

Fields

binding :: Word32

binding is the binding number of this entry and corresponds to a resource of the same binding number in the shader stages.

descriptorType :: DescriptorType

descriptorType is a DescriptorType specifying which type of resource descriptors are used for this binding.

descriptorCount :: Word32

descriptorCount is the number of descriptors contained in the binding, accessed in a shader as an array , except if descriptorType is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT in which case descriptorCount is the size in bytes of the inline uniform block . If descriptorCount is zero this binding entry is reserved and the resource must not be accessed from any stage via this binding within any pipeline using the set layout.

stageFlags :: ShaderStageFlags

stageFlags member is a bitmask of ShaderStageFlagBits specifying which pipeline shader stages can access a resource for this binding. SHADER_STAGE_ALL is a shorthand specifying that all defined shader stages, including any additional stages defined by extensions, can access the resource.

If a shader stage is not included in stageFlags, then a resource must not be accessed from that stage via this binding within any pipeline using the set layout. Other than input attachments which are limited to the fragment shader, there are no limitations on what combinations of stages can use a descriptor binding, and in particular a binding can be used by both graphics stages and the compute stage.

immutableSamplers :: Vector Sampler

data DescriptorSetLayoutCreateInfo (es :: [Type])

VkDescriptorSetLayoutCreateInfo - Structure specifying parameters of a newly created descriptor set layout

Valid Usage

The DescriptorSetLayoutBinding::binding members of the elements of the pBindings array must each have different values

If flags contains DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR, then all elements of pBindings must not have a descriptorType of DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC

If flags contains DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR, then all elements of pBindings must not have a descriptorType of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT

If flags contains DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR, then the total number of elements of all bindings must be less than or equal to PhysicalDevicePushDescriptorPropertiesKHR::maxPushDescriptors

If any binding has the DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT bit set, flags must include DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT

If any binding has the DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT bit set, then all bindings must not have descriptorType of DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO

pNext must be NULL or a pointer to a valid instance of DescriptorSetLayoutBindingFlagsCreateInfo

The sType value of each struct in the pNext chain must be unique

flags must be a valid combination of DescriptorSetLayoutCreateFlagBits values

If bindingCount is not 0, pBindings must be a valid pointer to an array of bindingCount valid DescriptorSetLayoutBinding structures

See Also

DescriptorSetLayoutBinding, DescriptorSetLayoutCreateFlags, StructureType, createDescriptorSetLayout, getDescriptorSetLayoutSupport, getDescriptorSetLayoutSupportKHR

Constructors

DescriptorSetLayoutCreateInfo

Fields

next :: Chain es

pNext is NULL or a pointer to a structure extending this structure.

flags :: DescriptorSetLayoutCreateFlags

flags is a bitmask of DescriptorSetLayoutCreateFlagBits specifying options for descriptor set layout creation.

bindings :: Vector DescriptorSetLayoutBinding

pBindings is a pointer to an array of DescriptorSetLayoutBinding structures.

data DescriptorPoolSize

VkDescriptorPoolSize - Structure specifying descriptor pool size

Description

Note

When creating a descriptor pool that will contain descriptors for combined image samplers of multi-planar formats, an application needs to account for non-trivial descriptor consumption when choosing the descriptorCount value, as indicated by SamplerYcbcrConversionImageFormatProperties::combinedImageSamplerDescriptorCount.

Valid Usage

descriptorCount must be greater than 0

If type is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then descriptorCount must be a multiple of 4

Valid Usage (Implicit)

type must be a valid DescriptorType value

See Also

DescriptorPoolCreateInfo, DescriptorType

Constructors

DescriptorPoolSize

Fields

type' :: DescriptorType

type is the type of descriptor.

descriptorCount :: Word32

descriptorCount is the number of descriptors of that type to allocate. If type is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then descriptorCount is the number of bytes to allocate for descriptors of this type.

data DescriptorPoolCreateInfo (es :: [Type])

VkDescriptorPoolCreateInfo - Structure specifying parameters of a newly created descriptor pool

Description

If multiple DescriptorPoolSize structures appear in the pPoolSizes array then the pool will be created with enough storage for the total number of descriptors of each type.

Fragmentation of a descriptor pool is possible and may lead to descriptor set allocation failures. A failure due to fragmentation is defined as failing a descriptor set allocation despite the sum of all outstanding descriptor set allocations from the pool plus the requested allocation requiring no more than the total number of descriptors requested at pool creation. Implementations provide certain guarantees of when fragmentation must not cause allocation failure, as described below.

If a descriptor pool has not had any descriptor sets freed since it was created or most recently reset then fragmentation must not cause an allocation failure (note that this is always the case for a pool created without the DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT bit set). Additionally, if all sets allocated from the pool since it was created or most recently reset use the same number of descriptors (of each type) and the requested allocation also uses that same number of descriptors (of each type), then fragmentation must not cause an allocation failure.

If an allocation failure occurs due to fragmentation, an application can create an additional descriptor pool to perform further descriptor set allocations.

If flags has the DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT bit set, descriptor pool creation may fail with the error ERROR_FRAGMENTATION if the total number of descriptors across all pools (including this one) created with this bit set exceeds maxUpdateAfterBindDescriptorsInAllPools, or if fragmentation of the underlying hardware resources occurs.

Valid Usage

maxSets must be greater than 0

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO

pNext must be NULL or a pointer to a valid instance of DescriptorPoolInlineUniformBlockCreateInfoEXT

The sType value of each struct in the pNext chain must be unique

flags must be a valid combination of DescriptorPoolCreateFlagBits values

pPoolSizes must be a valid pointer to an array of poolSizeCount valid DescriptorPoolSize structures

poolSizeCount must be greater than 0

See Also

DescriptorPoolCreateFlags, DescriptorPoolSize, StructureType, createDescriptorPool

Constructors

DescriptorPoolCreateInfo

Fields

next :: Chain es

pNext is NULL or a pointer to a structure extending this structure.

flags :: DescriptorPoolCreateFlags

flags is a bitmask of DescriptorPoolCreateFlagBits specifying certain supported operations on the pool.

maxSets :: Word32

maxSets is the maximum number of descriptor sets that can be allocated from the pool.

poolSizes :: Vector DescriptorPoolSize

pPoolSizes is a pointer to an array of DescriptorPoolSize structures, each containing a descriptor type and number of descriptors of that type to be allocated in the pool.

data DescriptorSetAllocateInfo (es :: [Type])

VkDescriptorSetAllocateInfo - Structure specifying the allocation parameters for descriptor sets

Valid Usage

Each element of pSetLayouts must not have been created with DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR set

If any element of pSetLayouts was created with the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set, descriptorPool must have been created with the DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT flag set

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO

pNext must be NULL or a pointer to a valid instance of DescriptorSetVariableDescriptorCountAllocateInfo

The sType value of each struct in the pNext chain must be unique

descriptorPool must be a valid DescriptorPool handle

pSetLayouts must be a valid pointer to an array of descriptorSetCount valid DescriptorSetLayout handles

descriptorSetCount must be greater than 0

Both of descriptorPool, and the elements of pSetLayouts must have been created, allocated, or retrieved from the same Device

See Also

DescriptorPool, DescriptorSetLayout, StructureType, allocateDescriptorSets

Constructors

DescriptorSetAllocateInfo

Fields

next :: Chain es

pNext is NULL or a pointer to a structure extending this structure.

descriptorPool :: DescriptorPool

descriptorPool is the pool which the sets will be allocated from.

setLayouts :: Vector DescriptorSetLayout

pSetLayouts is a pointer to an array of descriptor set layouts, with each member specifying how the corresponding descriptor set is allocated.

newtype DescriptorSet

VkDescriptorSet - Opaque handle to a descriptor set object

See Also

CopyDescriptorSet, WriteDescriptorSet, allocateDescriptorSets, cmdBindDescriptorSets, freeDescriptorSets, updateDescriptorSetWithTemplate, updateDescriptorSetWithTemplateKHR

Constructors

DescriptorSet Word64

newtype DescriptorSetLayout

VkDescriptorSetLayout - Opaque handle to a descriptor set layout object

See Also

DescriptorSetAllocateInfo, DescriptorUpdateTemplateCreateInfo, PipelineLayoutCreateInfo, createDescriptorSetLayout, destroyDescriptorSetLayout

Constructors

DescriptorSetLayout Word64

newtype DescriptorPool

VkDescriptorPool - Opaque handle to a descriptor pool object

See Also

DescriptorSetAllocateInfo, createDescriptorPool, destroyDescriptorPool, freeDescriptorSets, resetDescriptorPool

Constructors

DescriptorPool Word64

newtype DescriptorPoolResetFlags

VkDescriptorPoolResetFlags - Reserved for future use

Description

DescriptorPoolResetFlags is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

resetDescriptorPool

Constructors

DescriptorPoolResetFlags Flags

newtype DescriptorType

VkDescriptorType - Specifies the type of a descriptor in a descriptor set

Description

DESCRIPTOR_TYPE_SAMPLER specifies a sampler descriptor.

DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER specifies a combined image sampler descriptor.

DESCRIPTOR_TYPE_SAMPLED_IMAGE specifies a sampled image descriptor.

DESCRIPTOR_TYPE_STORAGE_IMAGE specifies a storage image descriptor.

DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER specifies a uniform texel buffer descriptor.

DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER specifies a storage texel buffer descriptor.

DESCRIPTOR_TYPE_UNIFORM_BUFFER specifies a uniform buffer descriptor.

DESCRIPTOR_TYPE_STORAGE_BUFFER specifies a storage buffer descriptor.

DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC specifies a dynamic uniform buffer descriptor.

DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC specifies a dynamic storage buffer descriptor.

DESCRIPTOR_TYPE_INPUT_ATTACHMENT specifies an input attachment descriptor.

DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT specifies an inline uniform block.

When a descriptor set is updated via elements of WriteDescriptorSet, members of pImageInfo, pBufferInfo and pTexelBufferView are only accessed by the implementation when they correspond to descriptor type being defined - otherwise they are ignored. The members accessed are as follows for each descriptor type:

For DESCRIPTOR_TYPE_SAMPLER, only the sampler member of each element of WriteDescriptorSet::pImageInfo is accessed.

For DESCRIPTOR_TYPE_SAMPLED_IMAGE, DESCRIPTOR_TYPE_STORAGE_IMAGE, or DESCRIPTOR_TYPE_INPUT_ATTACHMENT, only the imageView and imageLayout members of each element of WriteDescriptorSet::pImageInfo are accessed.

For DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, all members of each element of WriteDescriptorSet::pImageInfo are accessed.

For DESCRIPTOR_TYPE_UNIFORM_BUFFER, DESCRIPTOR_TYPE_STORAGE_BUFFER, DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC, all members of each element of WriteDescriptorSet::pBufferInfo are accessed.

For DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER or DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER, each element of WriteDescriptorSet::pTexelBufferView is accessed.

When updating descriptors with a descriptorType of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT, none of the pImageInfo, pBufferInfo, or pTexelBufferView members are accessed, instead the source data of the descriptor update operation is taken from the WriteDescriptorSetInlineUniformBlockEXT structure in the pNext chain of WriteDescriptorSet. When updating descriptors with a descriptorType of DESCRIPTOR_TYPE_ACCELERATION_STRUCTURE_KHR, none of the pImageInfo, pBufferInfo, or pTexelBufferView members are accessed, instead the source data of the descriptor update operation is taken from the WriteDescriptorSetAccelerationStructureKHR structure in the pNext chain of WriteDescriptorSet.

See Also

DescriptorPoolSize, DescriptorSetLayoutBinding, DescriptorUpdateTemplateEntry, ImageViewHandleInfoNVX, WriteDescriptorSet

Constructors

DescriptorType Int32

Bundled Patterns

pattern DESCRIPTOR_TYPE_SAMPLER :: DescriptorType

pattern DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER :: DescriptorType

pattern DESCRIPTOR_TYPE_SAMPLED_IMAGE :: DescriptorType

pattern DESCRIPTOR_TYPE_STORAGE_IMAGE :: DescriptorType

pattern DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER :: DescriptorType

pattern DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER :: DescriptorType

pattern DESCRIPTOR_TYPE_UNIFORM_BUFFER :: DescriptorType

pattern DESCRIPTOR_TYPE_STORAGE_BUFFER :: DescriptorType

pattern DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC :: DescriptorType

pattern DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC :: DescriptorType

pattern DESCRIPTOR_TYPE_INPUT_ATTACHMENT :: DescriptorType

pattern DESCRIPTOR_TYPE_ACCELERATION_STRUCTURE_KHR :: DescriptorType

pattern DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT :: DescriptorType

newtype DescriptorPoolCreateFlagBits

VkDescriptorPoolCreateFlagBits - Bitmask specifying certain supported operations on a descriptor pool

See Also

DescriptorPoolCreateFlags

Constructors

DescriptorPoolCreateFlagBits Flags

Bundled Patterns

pattern DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT :: DescriptorPoolCreateFlagBits
DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT specifies that descriptor sets can return their individual allocations to the pool, i.e. all of allocateDescriptorSets, freeDescriptorSets, and resetDescriptorPool are allowed. Otherwise, descriptor sets allocated from the pool must not be individually freed back to the pool, i.e. only allocateDescriptorSets and resetDescriptorPool are allowed.

pattern DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT :: DescriptorPoolCreateFlagBits
DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT specifies that descriptor sets allocated from this pool can include bindings with the DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT bit set. It is valid to allocate descriptor sets that have bindings that do not set the DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT bit from a pool that has DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT set.

type DescriptorPoolCreateFlags = DescriptorPoolCreateFlagBits

newtype DescriptorSetLayoutCreateFlagBits

VkDescriptorSetLayoutCreateFlagBits - Bitmask specifying descriptor set layout properties

See Also

DescriptorSetLayoutCreateFlags

Constructors

DescriptorSetLayoutCreateFlagBits Flags

Bundled Patterns

pattern DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR :: DescriptorSetLayoutCreateFlagBits
DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR specifies that descriptor sets must not be allocated using this layout, and descriptors are instead pushed by cmdPushDescriptorSetKHR.

pattern DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT :: DescriptorSetLayoutCreateFlagBits
DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT specifies that descriptor sets using this layout must be allocated from a descriptor pool created with the DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT bit set. Descriptor set layouts created with this bit set have alternate limits for the maximum number of descriptors per-stage and per-pipeline layout. The non-UpdateAfterBind limits only count descriptors in sets created without this flag. The UpdateAfterBind limits count all descriptors, but the limits may be higher than the non-UpdateAfterBind limits.

type DescriptorSetLayoutCreateFlags = DescriptorSetLayoutCreateFlagBits