Vulkan.Core10.Enums.AttachmentStoreOp

Documentation

newtype AttachmentStoreOp

VkAttachmentStoreOp - Specify how contents of an attachment are treated at the end of a subpass

Description

Note

ATTACHMENT_STORE_OP_DONT_CARE can cause contents generated during previous render passes to be discarded before reaching memory, even if no write to the attachment occurs during the current render pass.

See Also

AttachmentDescription, AttachmentDescription2

Constructors

AttachmentStoreOp Int32

Bundled Patterns

pattern ATTACHMENT_STORE_OP_STORE :: AttachmentStoreOp
ATTACHMENT_STORE_OP_STORE specifies the contents generated during the render pass and within the render area are written to memory. For attachments with a depth/stencil format, this uses the access type ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT. For attachments with a color format, this uses the access type ACCESS_COLOR_ATTACHMENT_WRITE_BIT.

pattern ATTACHMENT_STORE_OP_DONT_CARE :: AttachmentStoreOp
ATTACHMENT_STORE_OP_DONT_CARE specifies the contents within the render area are not needed after rendering, and may be discarded; the contents of the attachment will be undefined inside the render area. For attachments with a depth/stencil format, this uses the access type ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT. For attachments with a color format, this uses the access type ACCESS_COLOR_ATTACHMENT_WRITE_BIT.

pattern ATTACHMENT_STORE_OP_NONE_QCOM :: AttachmentStoreOp
ATTACHMENT_STORE_OP_NONE_QCOM specifies that the contents within the render area were not written during rendering, and may not be written to memory. If the attachment was written to during the renderpass, the contents of the attachment will be undefined inside the render area.