Vulkan.Core10.Enums.BorderColor

Documentation

newtype BorderColor

VkBorderColor - Specify border color used for texture lookups

Description

These colors are described in detail in Texel Replacement.

See Also

SamplerCreateInfo

Constructors

BorderColor Int32

Bundled Patterns

pattern BORDER_COLOR_FLOAT_TRANSPARENT_BLACK :: BorderColor
BORDER_COLOR_FLOAT_TRANSPARENT_BLACK specifies a transparent, floating-point format, black color.

pattern BORDER_COLOR_INT_TRANSPARENT_BLACK :: BorderColor
BORDER_COLOR_INT_TRANSPARENT_BLACK specifies a transparent, integer format, black color.

pattern BORDER_COLOR_FLOAT_OPAQUE_BLACK :: BorderColor
BORDER_COLOR_FLOAT_OPAQUE_BLACK specifies an opaque, floating-point format, black color.

pattern BORDER_COLOR_INT_OPAQUE_BLACK :: BorderColor
BORDER_COLOR_INT_OPAQUE_BLACK specifies an opaque, integer format, black color.

pattern BORDER_COLOR_FLOAT_OPAQUE_WHITE :: BorderColor
BORDER_COLOR_FLOAT_OPAQUE_WHITE specifies an opaque, floating-point format, white color.

pattern BORDER_COLOR_INT_OPAQUE_WHITE :: BorderColor
BORDER_COLOR_INT_OPAQUE_WHITE specifies an opaque, integer format, white color.

pattern BORDER_COLOR_INT_CUSTOM_EXT :: BorderColor
BORDER_COLOR_INT_CUSTOM_EXT indicates that a SamplerCustomBorderColorCreateInfoEXT structure is present in the SamplerCreateInfo::pNext chain which contains the color data in integer format.

pattern BORDER_COLOR_FLOAT_CUSTOM_EXT :: BorderColor
BORDER_COLOR_FLOAT_CUSTOM_EXT indicates that a SamplerCustomBorderColorCreateInfoEXT structure is present in the SamplerCreateInfo::pNext chain which contains the color data in floating-point format.