Vulkan.Core10.Enums.BufferCreateFlagBits

Documentation

type BufferCreateFlags = BufferCreateFlagBits

newtype BufferCreateFlagBits

VkBufferCreateFlagBits - Bitmask specifying additional parameters of a buffer

Description

See Sparse Resource Features and Physical Device Features for details of the sparse memory features supported on a device.

See Also

BufferCreateFlags

Constructors

BufferCreateFlagBits Flags

Bundled Patterns

pattern BUFFER_CREATE_SPARSE_BINDING_BIT :: BufferCreateFlagBits
BUFFER_CREATE_SPARSE_BINDING_BIT specifies that the buffer will be backed using sparse memory binding.

pattern BUFFER_CREATE_SPARSE_RESIDENCY_BIT :: BufferCreateFlagBits
BUFFER_CREATE_SPARSE_RESIDENCY_BIT specifies that the buffer can be partially backed using sparse memory binding. Buffers created with this flag must also be created with the BUFFER_CREATE_SPARSE_BINDING_BIT flag.

pattern BUFFER_CREATE_SPARSE_ALIASED_BIT :: BufferCreateFlagBits
BUFFER_CREATE_SPARSE_ALIASED_BIT specifies that the buffer will be backed using sparse memory binding with memory ranges that might also simultaneously be backing another buffer (or another portion of the same buffer). Buffers created with this flag must also be created with the BUFFER_CREATE_SPARSE_BINDING_BIT flag.

pattern BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT :: BufferCreateFlagBits
BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT specifies that the buffer’s address can be saved and reused on a subsequent run (e.g. for trace capture and replay), see BufferOpaqueCaptureAddressCreateInfo for more detail.

pattern BUFFER_CREATE_PROTECTED_BIT :: BufferCreateFlagBits
BUFFER_CREATE_PROTECTED_BIT specifies that the buffer is a protected buffer.