Vulkan.Core10.Enums.ColorComponentFlagBits

Documentation

newtype ColorComponentFlagBits

VkColorComponentFlagBits - Bitmask controlling which components are written to the framebuffer

Description

The color write mask operation is applied regardless of whether blending is enabled.

See Also

ColorComponentFlags

Constructors

ColorComponentFlagBits Flags

Bundled Patterns

pattern COLOR_COMPONENT_R_BIT :: ColorComponentFlagBits
COLOR_COMPONENT_R_BIT specifies that the R value is written to the color attachment for the appropriate sample. Otherwise, the value in memory is unmodified.

pattern COLOR_COMPONENT_G_BIT :: ColorComponentFlagBits
COLOR_COMPONENT_G_BIT specifies that the G value is written to the color attachment for the appropriate sample. Otherwise, the value in memory is unmodified.

pattern COLOR_COMPONENT_B_BIT :: ColorComponentFlagBits
COLOR_COMPONENT_B_BIT specifies that the B value is written to the color attachment for the appropriate sample. Otherwise, the value in memory is unmodified.

pattern COLOR_COMPONENT_A_BIT :: ColorComponentFlagBits
COLOR_COMPONENT_A_BIT specifies that the A value is written to the color attachment for the appropriate sample. Otherwise, the value in memory is unmodified.

type ColorComponentFlags = ColorComponentFlagBits