Vulkan.Core10.Enums.CommandPoolCreateFlagBits

Documentation

newtype CommandPoolCreateFlagBits

VkCommandPoolCreateFlagBits - Bitmask specifying usage behavior for a command pool

See Also

CommandPoolCreateFlags

Constructors

CommandPoolCreateFlagBits Flags

Bundled Patterns

pattern COMMAND_POOL_CREATE_TRANSIENT_BIT :: CommandPoolCreateFlagBits
COMMAND_POOL_CREATE_TRANSIENT_BIT specifies that command buffers allocated from the pool will be short-lived, meaning that they will be reset or freed in a relatively short timeframe. This flag may be used by the implementation to control memory allocation behavior within the pool.

pattern COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT :: CommandPoolCreateFlagBits
COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT allows any command buffer allocated from a pool to be individually reset to the initial state; either by calling resetCommandBuffer, or via the implicit reset when calling beginCommandBuffer. If this flag is not set on a pool, then resetCommandBuffer must not be called for any command buffer allocated from that pool.

pattern COMMAND_POOL_CREATE_PROTECTED_BIT :: CommandPoolCreateFlagBits
COMMAND_POOL_CREATE_PROTECTED_BIT specifies that command buffers allocated from the pool are protected command buffers.

type CommandPoolCreateFlags = CommandPoolCreateFlagBits