Vulkan.Core10.Enums.DescriptorType

Documentation

newtype DescriptorType

VkDescriptorType - Specifies the type of a descriptor in a descriptor set

Description

DESCRIPTOR_TYPE_SAMPLER specifies a sampler descriptor.

DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER specifies a combined image sampler descriptor.

DESCRIPTOR_TYPE_SAMPLED_IMAGE specifies a sampled image descriptor.

DESCRIPTOR_TYPE_STORAGE_IMAGE specifies a storage image descriptor.

DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER specifies a uniform texel buffer descriptor.

DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER specifies a storage texel buffer descriptor.

DESCRIPTOR_TYPE_UNIFORM_BUFFER specifies a uniform buffer descriptor.

DESCRIPTOR_TYPE_STORAGE_BUFFER specifies a storage buffer descriptor.

DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC specifies a dynamic uniform buffer descriptor.

DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC specifies a dynamic storage buffer descriptor.

DESCRIPTOR_TYPE_INPUT_ATTACHMENT specifies an input attachment descriptor.

DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT specifies an inline uniform block.

When a descriptor set is updated via elements of WriteDescriptorSet, members of pImageInfo, pBufferInfo and pTexelBufferView are only accessed by the implementation when they correspond to descriptor type being defined - otherwise they are ignored. The members accessed are as follows for each descriptor type:

For DESCRIPTOR_TYPE_SAMPLER, only the sampler member of each element of WriteDescriptorSet::pImageInfo is accessed.

For DESCRIPTOR_TYPE_SAMPLED_IMAGE, DESCRIPTOR_TYPE_STORAGE_IMAGE, or DESCRIPTOR_TYPE_INPUT_ATTACHMENT, only the imageView and imageLayout members of each element of WriteDescriptorSet::pImageInfo are accessed.

For DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, all members of each element of WriteDescriptorSet::pImageInfo are accessed.

For DESCRIPTOR_TYPE_UNIFORM_BUFFER, DESCRIPTOR_TYPE_STORAGE_BUFFER, DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC, all members of each element of WriteDescriptorSet::pBufferInfo are accessed.

For DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER or DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER, each element of WriteDescriptorSet::pTexelBufferView is accessed.

When updating descriptors with a descriptorType of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT, none of the pImageInfo, pBufferInfo, or pTexelBufferView members are accessed, instead the source data of the descriptor update operation is taken from the WriteDescriptorSetInlineUniformBlockEXT structure in the pNext chain of WriteDescriptorSet. When updating descriptors with a descriptorType of DESCRIPTOR_TYPE_ACCELERATION_STRUCTURE_KHR, none of the pImageInfo, pBufferInfo, or pTexelBufferView members are accessed, instead the source data of the descriptor update operation is taken from the WriteDescriptorSetAccelerationStructureKHR structure in the pNext chain of WriteDescriptorSet.

See Also

DescriptorPoolSize, DescriptorSetLayoutBinding, DescriptorUpdateTemplateEntry, ImageViewHandleInfoNVX, WriteDescriptorSet

Constructors

DescriptorType Int32

Bundled Patterns

pattern DESCRIPTOR_TYPE_SAMPLER :: DescriptorType

pattern DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER :: DescriptorType

pattern DESCRIPTOR_TYPE_SAMPLED_IMAGE :: DescriptorType

pattern DESCRIPTOR_TYPE_STORAGE_IMAGE :: DescriptorType

pattern DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER :: DescriptorType

pattern DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER :: DescriptorType

pattern DESCRIPTOR_TYPE_UNIFORM_BUFFER :: DescriptorType

pattern DESCRIPTOR_TYPE_STORAGE_BUFFER :: DescriptorType

pattern DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC :: DescriptorType

pattern DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC :: DescriptorType

pattern DESCRIPTOR_TYPE_INPUT_ATTACHMENT :: DescriptorType

pattern DESCRIPTOR_TYPE_ACCELERATION_STRUCTURE_KHR :: DescriptorType

pattern DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT :: DescriptorType