Vulkan.Core10.Enums.DynamicState

Documentation

newtype DynamicState

VkDynamicState - Indicate which dynamic state is taken from dynamic state commands

See Also

PipelineDynamicStateCreateInfo

Constructors

DynamicState Int32

Bundled Patterns

pattern DYNAMIC_STATE_VIEWPORT :: DynamicState
DYNAMIC_STATE_VIEWPORT specifies that the pViewports state in PipelineViewportStateCreateInfo will be ignored and must be set dynamically with cmdSetViewport before any draw commands. The number of viewports used by a pipeline is still specified by the viewportCount member of PipelineViewportStateCreateInfo.

pattern DYNAMIC_STATE_SCISSOR :: DynamicState
DYNAMIC_STATE_SCISSOR specifies that the pScissors state in PipelineViewportStateCreateInfo will be ignored and must be set dynamically with cmdSetScissor before any draw commands. The number of scissor rectangles used by a pipeline is still specified by the scissorCount member of PipelineViewportStateCreateInfo.

pattern DYNAMIC_STATE_LINE_WIDTH :: DynamicState
DYNAMIC_STATE_LINE_WIDTH specifies that the lineWidth state in PipelineRasterizationStateCreateInfo will be ignored and must be set dynamically with cmdSetLineWidth before any draw commands that generate line primitives for the rasterizer.

pattern DYNAMIC_STATE_DEPTH_BIAS :: DynamicState
DYNAMIC_STATE_DEPTH_BIAS specifies that the depthBiasConstantFactor, depthBiasClamp and depthBiasSlopeFactor states in PipelineRasterizationStateCreateInfo will be ignored and must be set dynamically with cmdSetDepthBias before any draws are performed with depthBiasEnable in PipelineRasterizationStateCreateInfo set to TRUE.

pattern DYNAMIC_STATE_BLEND_CONSTANTS :: DynamicState
DYNAMIC_STATE_BLEND_CONSTANTS specifies that the blendConstants state in PipelineColorBlendStateCreateInfo will be ignored and must be set dynamically with cmdSetBlendConstants before any draws are performed with a pipeline state with PipelineColorBlendAttachmentState member blendEnable set to TRUE and any of the blend functions using a constant blend color.

pattern DYNAMIC_STATE_DEPTH_BOUNDS :: DynamicState
DYNAMIC_STATE_DEPTH_BOUNDS specifies that the minDepthBounds and maxDepthBounds states of PipelineDepthStencilStateCreateInfo will be ignored and must be set dynamically with cmdSetDepthBounds before any draws are performed with a pipeline state with PipelineDepthStencilStateCreateInfo member depthBoundsTestEnable set to TRUE.

pattern DYNAMIC_STATE_STENCIL_COMPARE_MASK :: DynamicState
DYNAMIC_STATE_STENCIL_COMPARE_MASK specifies that the compareMask state in PipelineDepthStencilStateCreateInfo for both front and back will be ignored and must be set dynamically with cmdSetStencilCompareMask before any draws are performed with a pipeline state with PipelineDepthStencilStateCreateInfo member stencilTestEnable set to TRUE

pattern DYNAMIC_STATE_STENCIL_WRITE_MASK :: DynamicState
DYNAMIC_STATE_STENCIL_WRITE_MASK specifies that the writeMask state in PipelineDepthStencilStateCreateInfo for both front and back will be ignored and must be set dynamically with cmdSetStencilWriteMask before any draws are performed with a pipeline state with PipelineDepthStencilStateCreateInfo member stencilTestEnable set to TRUE

pattern DYNAMIC_STATE_STENCIL_REFERENCE :: DynamicState
DYNAMIC_STATE_STENCIL_REFERENCE specifies that the reference state in PipelineDepthStencilStateCreateInfo for both front and back will be ignored and must be set dynamically with cmdSetStencilReference before any draws are performed with a pipeline state with PipelineDepthStencilStateCreateInfo member stencilTestEnable set to TRUE

pattern DYNAMIC_STATE_STENCIL_OP_EXT :: DynamicState
DYNAMIC_STATE_STENCIL_OP_EXT specifies that the failOp, passOp, depthFailOp, and compareOp states in PipelineDepthStencilStateCreateInfo for both front and back will be ignored and must be set dynamically with cmdSetStencilOpEXT before any draws are performed with a pipeline state with PipelineDepthStencilStateCreateInfo member stencilTestEnable set to TRUE

pattern DYNAMIC_STATE_STENCIL_TEST_ENABLE_EXT :: DynamicState
DYNAMIC_STATE_STENCIL_TEST_ENABLE_EXT specifies that the stencilTestEnable state in PipelineDepthStencilStateCreateInfo will be ignored and must be set dynamically with cmdSetStencilTestEnableEXT before any draw call.

pattern DYNAMIC_STATE_DEPTH_BOUNDS_TEST_ENABLE_EXT :: DynamicState
DYNAMIC_STATE_DEPTH_BOUNDS_TEST_ENABLE_EXT specifies that the depthBoundsTestEnable state in PipelineDepthStencilStateCreateInfo will be ignored and must be set dynamically with cmdSetDepthBoundsTestEnableEXT before any draw call.

pattern DYNAMIC_STATE_DEPTH_COMPARE_OP_EXT :: DynamicState
DYNAMIC_STATE_DEPTH_COMPARE_OP_EXT specifies that the depthCompareOp state in PipelineDepthStencilStateCreateInfo will be ignored and must be set dynamically with cmdSetDepthCompareOpEXT before any draw call.

pattern DYNAMIC_STATE_DEPTH_WRITE_ENABLE_EXT :: DynamicState
DYNAMIC_STATE_DEPTH_WRITE_ENABLE_EXT specifies that the depthWriteEnable state in PipelineDepthStencilStateCreateInfo will be ignored and must be set dynamically with cmdSetDepthWriteEnableEXT before any draw call.

pattern DYNAMIC_STATE_DEPTH_TEST_ENABLE_EXT :: DynamicState
DYNAMIC_STATE_DEPTH_TEST_ENABLE_EXT specifies that the depthTestEnable state in PipelineDepthStencilStateCreateInfo will be ignored and must be set dynamically with cmdSetDepthTestEnableEXT before any draw call.

pattern DYNAMIC_STATE_VERTEX_INPUT_BINDING_STRIDE_EXT :: DynamicState
DYNAMIC_STATE_VERTEX_INPUT_BINDING_STRIDE_EXT specifies that the stride state in VertexInputBindingDescription will be ignored and must be set dynamically with cmdBindVertexBuffers2EXT before any draw call.

pattern DYNAMIC_STATE_SCISSOR_WITH_COUNT_EXT :: DynamicState
DYNAMIC_STATE_SCISSOR_WITH_COUNT_EXT specifies that the scissorCount and pScissors state in PipelineViewportStateCreateInfo will be ignored and must be set dynamically with cmdSetScissorWithCountEXT before any draw call.

pattern DYNAMIC_STATE_VIEWPORT_WITH_COUNT_EXT :: DynamicState
DYNAMIC_STATE_VIEWPORT_WITH_COUNT_EXT specifies that the viewportCount and pViewports state in PipelineViewportStateCreateInfo will be ignored and must be set dynamically with cmdSetViewportWithCountEXT before any draw call.

pattern DYNAMIC_STATE_PRIMITIVE_TOPOLOGY_EXT :: DynamicState
DYNAMIC_STATE_PRIMITIVE_TOPOLOGY_EXT specifies that the topology state in PipelineInputAssemblyStateCreateInfo only specifies the topology class, and the specific topology order and adjacency must be set dynamically with cmdSetPrimitiveTopologyEXT before any draw commands.

pattern DYNAMIC_STATE_FRONT_FACE_EXT :: DynamicState
DYNAMIC_STATE_FRONT_FACE_EXT specifies that the frontFace state in PipelineRasterizationStateCreateInfo will be ignored and must be set dynamically with cmdSetFrontFaceEXT before any draw commands.

pattern DYNAMIC_STATE_CULL_MODE_EXT :: DynamicState
DYNAMIC_STATE_CULL_MODE_EXT specifies that the cullMode state in PipelineRasterizationStateCreateInfo will be ignored and must be set dynamically with cmdSetCullModeEXT before any draw commands.

pattern DYNAMIC_STATE_LINE_STIPPLE_EXT :: DynamicState
DYNAMIC_STATE_LINE_STIPPLE_EXT specifies that the lineStippleFactor and lineStipplePattern state in PipelineRasterizationLineStateCreateInfoEXT will be ignored and must be set dynamically with cmdSetLineStippleEXT before any draws are performed with a pipeline state with PipelineRasterizationLineStateCreateInfoEXT member stippledLineEnable set to TRUE.

pattern DYNAMIC_STATE_FRAGMENT_SHADING_RATE_KHR :: DynamicState
DYNAMIC_STATE_FRAGMENT_SHADING_RATE_KHR specifies that state in PipelineFragmentShadingRateStateCreateInfoKHR and PipelineFragmentShadingRateEnumStateCreateInfoNV will be ignored and must be set dynamically with cmdSetFragmentShadingRateKHR or cmdSetFragmentShadingRateEnumNV before any draw commands.

pattern DYNAMIC_STATE_EXCLUSIVE_SCISSOR_NV :: DynamicState
DYNAMIC_STATE_EXCLUSIVE_SCISSOR_NV specifies that the pExclusiveScissors state in PipelineViewportExclusiveScissorStateCreateInfoNV will be ignored and must be set dynamically with cmdSetExclusiveScissorNV before any draw commands. The number of exclusive scissor rectangles used by a pipeline is still specified by the exclusiveScissorCount member of PipelineViewportExclusiveScissorStateCreateInfoNV.

pattern DYNAMIC_STATE_VIEWPORT_COARSE_SAMPLE_ORDER_NV :: DynamicState
DYNAMIC_STATE_VIEWPORT_COARSE_SAMPLE_ORDER_NV specifies that the coarse sample order state in PipelineViewportCoarseSampleOrderStateCreateInfoNV will be ignored and must be set dynamically with cmdSetCoarseSampleOrderNV before any draw commands.

pattern DYNAMIC_STATE_VIEWPORT_SHADING_RATE_PALETTE_NV :: DynamicState
DYNAMIC_STATE_VIEWPORT_SHADING_RATE_PALETTE_NV specifies that the pShadingRatePalettes state in PipelineViewportShadingRateImageStateCreateInfoNV will be ignored and must be set dynamically with cmdSetViewportShadingRatePaletteNV before any draw commands.

pattern DYNAMIC_STATE_RAY_TRACING_PIPELINE_STACK_SIZE_KHR :: DynamicState
DYNAMIC_STATE_RAY_TRACING_PIPELINE_STACK_SIZE_KHR specifies that the default stack size computation for the pipeline will be ignored and must be set dynamically with cmdSetRayTracingPipelineStackSizeKHR before any ray tracing calls are performed.

pattern DYNAMIC_STATE_SAMPLE_LOCATIONS_EXT :: DynamicState
DYNAMIC_STATE_SAMPLE_LOCATIONS_EXT specifies that the sampleLocationsInfo state in PipelineSampleLocationsStateCreateInfoEXT will be ignored and must be set dynamically with cmdSetSampleLocationsEXT before any draw or clear commands. Enabling custom sample locations is still indicated by the sampleLocationsEnable member of PipelineSampleLocationsStateCreateInfoEXT.

pattern DYNAMIC_STATE_DISCARD_RECTANGLE_EXT :: DynamicState
DYNAMIC_STATE_DISCARD_RECTANGLE_EXT specifies that the pDiscardRectangles state in PipelineDiscardRectangleStateCreateInfoEXT will be ignored and must be set dynamically with cmdSetDiscardRectangleEXT before any draw or clear commands. The DiscardRectangleModeEXT and the number of active discard rectangles is still specified by the discardRectangleMode and discardRectangleCount members of PipelineDiscardRectangleStateCreateInfoEXT.

pattern DYNAMIC_STATE_VIEWPORT_W_SCALING_NV :: DynamicState
DYNAMIC_STATE_VIEWPORT_W_SCALING_NV specifies that the pViewportScalings state in PipelineViewportWScalingStateCreateInfoNV will be ignored and must be set dynamically with cmdSetViewportWScalingNV before any draws are performed with a pipeline state with PipelineViewportWScalingStateCreateInfoNV member viewportScalingEnable set to TRUE