Vulkan.Core10.Enums.ImageViewType

Documentation

newtype ImageViewType

VkImageViewType - Image view types

Description

The exact image view type is partially implicit, based on the image’s type and sample count, as well as the view creation parameters as described in the image view compatibility table for createImageView. This table also shows which SPIR-V OpTypeImage Dim and Arrayed parameters correspond to each image view type.

See Also

ImageViewCreateInfo, PhysicalDeviceImageViewImageFormatInfoEXT

Constructors

ImageViewType Int32

Bundled Patterns

pattern IMAGE_VIEW_TYPE_1D :: ImageViewType

pattern IMAGE_VIEW_TYPE_2D :: ImageViewType

pattern IMAGE_VIEW_TYPE_3D :: ImageViewType

pattern IMAGE_VIEW_TYPE_CUBE :: ImageViewType

pattern IMAGE_VIEW_TYPE_1D_ARRAY :: ImageViewType

pattern IMAGE_VIEW_TYPE_2D_ARRAY :: ImageViewType

pattern IMAGE_VIEW_TYPE_CUBE_ARRAY :: ImageViewType