Vulkan.Core10.Enums.PhysicalDeviceType

Documentation

newtype PhysicalDeviceType

VkPhysicalDeviceType - Supported physical device types

Description

The physical device type is advertised for informational purposes only, and does not directly affect the operation of the system. However, the device type may correlate with other advertised properties or capabilities of the system, such as how many memory heaps there are.

See Also

PhysicalDeviceProperties

Constructors

PhysicalDeviceType Int32

Bundled Patterns

pattern PHYSICAL_DEVICE_TYPE_OTHER :: PhysicalDeviceType
PHYSICAL_DEVICE_TYPE_OTHER - the device does not match any other available types.

pattern PHYSICAL_DEVICE_TYPE_INTEGRATED_GPU :: PhysicalDeviceType
PHYSICAL_DEVICE_TYPE_INTEGRATED_GPU - the device is typically one embedded in or tightly coupled with the host.

pattern PHYSICAL_DEVICE_TYPE_DISCRETE_GPU :: PhysicalDeviceType
PHYSICAL_DEVICE_TYPE_DISCRETE_GPU - the device is typically a separate processor connected to the host via an interlink.

pattern PHYSICAL_DEVICE_TYPE_VIRTUAL_GPU :: PhysicalDeviceType
PHYSICAL_DEVICE_TYPE_VIRTUAL_GPU - the device is typically a virtual node in a virtualization environment.

pattern PHYSICAL_DEVICE_TYPE_CPU :: PhysicalDeviceType
PHYSICAL_DEVICE_TYPE_CPU - the device is typically running on the same processors as the host.