Vulkan.Core10.Enums.PipelineStageFlagBits

Documentation

type PipelineStageFlags = PipelineStageFlagBits

newtype PipelineStageFlagBits

VkPipelineStageFlagBits - Bitmask specifying pipeline stages

See Also

CheckpointDataNV, PipelineStageFlags, cmdWriteBufferMarkerAMD, cmdWriteTimestamp

Constructors

PipelineStageFlagBits Flags

Bundled Patterns

pattern PIPELINE_STAGE_TOP_OF_PIPE_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_TOP_OF_PIPE_BIT is equivalent to PIPELINE_STAGE_ALL_COMMANDS_BIT with AccessFlags set to 0 when specified in the second synchronization scope, but specifies no stages in the first scope.

pattern PIPELINE_STAGE_DRAW_INDIRECT_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_DRAW_INDIRECT_BIT specifies the stage of the pipeline where Draw/DispatchIndirect/TraceRaysIndirect data structures are consumed. This stage also includes reading commands written by cmdExecuteGeneratedCommandsNV.

pattern PIPELINE_STAGE_VERTEX_INPUT_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_VERTEX_INPUT_BIT specifies the stage of the pipeline where vertex and index buffers are consumed.

pattern PIPELINE_STAGE_VERTEX_SHADER_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_VERTEX_SHADER_BIT specifies the vertex shader stage.

pattern PIPELINE_STAGE_TESSELLATION_CONTROL_SHADER_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_TESSELLATION_CONTROL_SHADER_BIT specifies the tessellation control shader stage.

pattern PIPELINE_STAGE_TESSELLATION_EVALUATION_SHADER_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_TESSELLATION_EVALUATION_SHADER_BIT specifies the tessellation evaluation shader stage.

pattern PIPELINE_STAGE_GEOMETRY_SHADER_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_GEOMETRY_SHADER_BIT specifies the geometry shader stage.

pattern PIPELINE_STAGE_FRAGMENT_SHADER_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_FRAGMENT_SHADER_BIT specifies the fragment shader stage.

pattern PIPELINE_STAGE_EARLY_FRAGMENT_TESTS_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_EARLY_FRAGMENT_TESTS_BIT specifies the stage of the pipeline where early fragment tests (depth and stencil tests before fragment shading) are performed. This stage also includes subpass load operations for framebuffer attachments with a depth/stencil format.

pattern PIPELINE_STAGE_LATE_FRAGMENT_TESTS_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_LATE_FRAGMENT_TESTS_BIT specifies the stage of the pipeline where late fragment tests (depth and stencil tests after fragment shading) are performed. This stage also includes subpass store operations for framebuffer attachments with a depth/stencil format.

pattern PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT specifies the stage of the pipeline after blending where the final color values are output from the pipeline. This stage also includes subpass load and store operations and multisample resolve operations for framebuffer attachments with a color or depth/stencil format.

pattern PIPELINE_STAGE_COMPUTE_SHADER_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_COMPUTE_SHADER_BIT specifies the execution of a compute shader.

pattern PIPELINE_STAGE_TRANSFER_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_TRANSFER_BIT specifies the following commands:

All copy commands, including cmdCopyQueryPoolResults

cmdBlitImage2KHR and cmdBlitImage

cmdResolveImage2KHR and cmdResolveImage

All clear commands, with the exception of cmdClearAttachments

pattern PIPELINE_STAGE_BOTTOM_OF_PIPE_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_BOTTOM_OF_PIPE_BIT is equivalent to PIPELINE_STAGE_ALL_COMMANDS_BIT with AccessFlags set to 0 when specified in the first synchronization scope, but specifies no stages in the second scope.

pattern PIPELINE_STAGE_HOST_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_HOST_BIT specifies a pseudo-stage indicating execution on the host of reads/writes of device memory. This stage is not invoked by any commands recorded in a command buffer.

pattern PIPELINE_STAGE_ALL_GRAPHICS_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_ALL_GRAPHICS_BIT specifies the execution of all graphics pipeline stages, and is equivalent to the logical OR of:

PIPELINE_STAGE_DRAW_INDIRECT_BIT

PIPELINE_STAGE_TASK_SHADER_BIT_NV

PIPELINE_STAGE_MESH_SHADER_BIT_NV

PIPELINE_STAGE_VERTEX_INPUT_BIT

PIPELINE_STAGE_VERTEX_SHADER_BIT

PIPELINE_STAGE_TESSELLATION_CONTROL_SHADER_BIT

PIPELINE_STAGE_TESSELLATION_EVALUATION_SHADER_BIT

PIPELINE_STAGE_GEOMETRY_SHADER_BIT

PIPELINE_STAGE_FRAGMENT_SHADER_BIT

PIPELINE_STAGE_EARLY_FRAGMENT_TESTS_BIT

PIPELINE_STAGE_LATE_FRAGMENT_TESTS_BIT

PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT

PIPELINE_STAGE_CONDITIONAL_RENDERING_BIT_EXT

PIPELINE_STAGE_TRANSFORM_FEEDBACK_BIT_EXT

PIPELINE_STAGE_FRAGMENT_SHADING_RATE_ATTACHMENT_BIT_KHR

PIPELINE_STAGE_FRAGMENT_DENSITY_PROCESS_BIT_EXT

pattern PIPELINE_STAGE_ALL_COMMANDS_BIT :: PipelineStageFlagBits
PIPELINE_STAGE_ALL_COMMANDS_BIT specifies all commands supported on the queue it is used with.

pattern PIPELINE_STAGE_COMMAND_PREPROCESS_BIT_NV :: PipelineStageFlagBits
PIPELINE_STAGE_COMMAND_PREPROCESS_BIT_NV specifies the stage of the pipeline where device-side preprocessing for generated commands via cmdPreprocessGeneratedCommandsNV is handled.

pattern PIPELINE_STAGE_FRAGMENT_DENSITY_PROCESS_BIT_EXT :: PipelineStageFlagBits
PIPELINE_STAGE_FRAGMENT_DENSITY_PROCESS_BIT_EXT specifies the stage of the pipeline where the fragment density map is read to generate the fragment areas.

pattern PIPELINE_STAGE_MESH_SHADER_BIT_NV :: PipelineStageFlagBits
PIPELINE_STAGE_MESH_SHADER_BIT_NV specifies the mesh shader stage.

pattern PIPELINE_STAGE_TASK_SHADER_BIT_NV :: PipelineStageFlagBits
PIPELINE_STAGE_TASK_SHADER_BIT_NV specifies the task shader stage.

pattern PIPELINE_STAGE_SHADING_RATE_IMAGE_BIT_NV :: PipelineStageFlagBits

pattern PIPELINE_STAGE_RAY_TRACING_SHADER_BIT_KHR :: PipelineStageFlagBits
PIPELINE_STAGE_RAY_TRACING_SHADER_BIT_KHR specifies the execution of the ray tracing shader stages, via cmdTraceRaysNV , cmdTraceRaysKHR, or cmdTraceRaysIndirectKHR

pattern PIPELINE_STAGE_ACCELERATION_STRUCTURE_BUILD_BIT_KHR :: PipelineStageFlagBits
PIPELINE_STAGE_ACCELERATION_STRUCTURE_BUILD_BIT_KHR specifies the execution of cmdBuildAccelerationStructureNV, cmdCopyAccelerationStructureNV, cmdWriteAccelerationStructuresPropertiesNV , cmdBuildAccelerationStructuresKHR, cmdBuildAccelerationStructuresIndirectKHR, cmdCopyAccelerationStructureKHR, cmdCopyAccelerationStructureToMemoryKHR, cmdCopyMemoryToAccelerationStructureKHR, and cmdWriteAccelerationStructuresPropertiesKHR.

pattern PIPELINE_STAGE_CONDITIONAL_RENDERING_BIT_EXT :: PipelineStageFlagBits
PIPELINE_STAGE_CONDITIONAL_RENDERING_BIT_EXT specifies the stage of the pipeline where the predicate of conditional rendering is consumed.

pattern PIPELINE_STAGE_TRANSFORM_FEEDBACK_BIT_EXT :: PipelineStageFlagBits
PIPELINE_STAGE_TRANSFORM_FEEDBACK_BIT_EXT specifies the stage of the pipeline where vertex attribute output values are written to the transform feedback buffers.