Vulkan.Core10.Enums.QueueFlagBits

Documentation

newtype QueueFlagBits

VkQueueFlagBits - Bitmask specifying capabilities of queues in a queue family

Description

QUEUE_GRAPHICS_BIT specifies that queues in this queue family support graphics operations.

QUEUE_COMPUTE_BIT specifies that queues in this queue family support compute operations.

QUEUE_TRANSFER_BIT specifies that queues in this queue family support transfer operations.

QUEUE_SPARSE_BINDING_BIT specifies that queues in this queue family support sparse memory management operations (see Sparse Resources). If any of the sparse resource features are enabled, then at least one queue family must support this bit.

if QUEUE_PROTECTED_BIT is set, then the queues in this queue family support the DEVICE_QUEUE_CREATE_PROTECTED_BIT bit. (see Protected Memory). If the protected memory physical device feature is supported, then at least one queue family of at least one physical device exposed by the implementation must support this bit.

If an implementation exposes any queue family that supports graphics operations, at least one queue family of at least one physical device exposed by the implementation must support both graphics and compute operations.

Furthermore, if the protected memory physical device feature is supported, then at least one queue family of at least one physical device exposed by the implementation must support graphics operations, compute operations, and protected memory operations.

Note

All commands that are allowed on a queue that supports transfer operations are also allowed on a queue that supports either graphics or compute operations. Thus, if the capabilities of a queue family include QUEUE_GRAPHICS_BIT or QUEUE_COMPUTE_BIT, then reporting the QUEUE_TRANSFER_BIT capability separately for that queue family is optional.

For further details see Queues.

See Also

QueueFlags

Constructors

QueueFlagBits Flags

Bundled Patterns

pattern QUEUE_GRAPHICS_BIT :: QueueFlagBits

pattern QUEUE_COMPUTE_BIT :: QueueFlagBits

pattern QUEUE_TRANSFER_BIT :: QueueFlagBits

pattern QUEUE_SPARSE_BINDING_BIT :: QueueFlagBits

pattern QUEUE_PROTECTED_BIT :: QueueFlagBits

type QueueFlags = QueueFlagBits