Vulkan.Core10.Enums.ShaderStageFlagBits

Documentation

newtype ShaderStageFlagBits

VkShaderStageFlagBits - Bitmask specifying a pipeline stage

Description

Note

SHADER_STAGE_ALL_GRAPHICS only includes the original five graphics stages included in Vulkan 1.0, and not any stages added by extensions. Thus, it may not have the desired effect in all cases.

See Also

PipelineShaderStageCreateInfo, ShaderStageFlags, getShaderInfoAMD

Constructors

ShaderStageFlagBits Flags

Bundled Patterns

pattern SHADER_STAGE_VERTEX_BIT :: ShaderStageFlagBits
SHADER_STAGE_VERTEX_BIT specifies the vertex stage.

pattern SHADER_STAGE_TESSELLATION_CONTROL_BIT :: ShaderStageFlagBits
SHADER_STAGE_TESSELLATION_CONTROL_BIT specifies the tessellation control stage.

pattern SHADER_STAGE_TESSELLATION_EVALUATION_BIT :: ShaderStageFlagBits
SHADER_STAGE_TESSELLATION_EVALUATION_BIT specifies the tessellation evaluation stage.

pattern SHADER_STAGE_GEOMETRY_BIT :: ShaderStageFlagBits
SHADER_STAGE_GEOMETRY_BIT specifies the geometry stage.

pattern SHADER_STAGE_FRAGMENT_BIT :: ShaderStageFlagBits
SHADER_STAGE_FRAGMENT_BIT specifies the fragment stage.

pattern SHADER_STAGE_COMPUTE_BIT :: ShaderStageFlagBits
SHADER_STAGE_COMPUTE_BIT specifies the compute stage.

pattern SHADER_STAGE_ALL_GRAPHICS :: ShaderStageFlagBits
SHADER_STAGE_ALL_GRAPHICS is a combination of bits used as shorthand to specify all graphics stages defined above (excluding the compute stage).

pattern SHADER_STAGE_ALL :: ShaderStageFlagBits
SHADER_STAGE_ALL is a combination of bits used as shorthand to specify all shader stages supported by the device, including all additional stages which are introduced by extensions.

pattern SHADER_STAGE_MESH_BIT_NV :: ShaderStageFlagBits
SHADER_STAGE_MESH_BIT_NV specifies the mesh stage.

pattern SHADER_STAGE_TASK_BIT_NV :: ShaderStageFlagBits
SHADER_STAGE_TASK_BIT_NV specifies the task stage.

pattern SHADER_STAGE_CALLABLE_BIT_KHR :: ShaderStageFlagBits
SHADER_STAGE_CALLABLE_BIT_KHR specifies the callable stage.

pattern SHADER_STAGE_INTERSECTION_BIT_KHR :: ShaderStageFlagBits
SHADER_STAGE_INTERSECTION_BIT_KHR specifies the intersection stage.

pattern SHADER_STAGE_MISS_BIT_KHR :: ShaderStageFlagBits
SHADER_STAGE_MISS_BIT_KHR specifies the miss stage.

pattern SHADER_STAGE_CLOSEST_HIT_BIT_KHR :: ShaderStageFlagBits
SHADER_STAGE_CLOSEST_HIT_BIT_KHR specifies the closest hit stage.

pattern SHADER_STAGE_ANY_HIT_BIT_KHR :: ShaderStageFlagBits
SHADER_STAGE_ANY_HIT_BIT_KHR specifies the any-hit stage.

pattern SHADER_STAGE_RAYGEN_BIT_KHR :: ShaderStageFlagBits
SHADER_STAGE_RAYGEN_BIT_KHR specifies the ray generation stage.

type ShaderStageFlags = ShaderStageFlagBits