Vulkan.Core10.Enums.StencilOp

Documentation

newtype StencilOp

VkStencilOp - Stencil comparison function

Description

For purposes of increment and decrement, the stencil bits are considered as an unsigned integer.

See Also

StencilOpState

Constructors

StencilOp Int32

Bundled Patterns

pattern STENCIL_OP_KEEP :: StencilOp
STENCIL_OP_KEEP keeps the current value.

pattern STENCIL_OP_ZERO :: StencilOp
STENCIL_OP_ZERO sets the value to 0.

pattern STENCIL_OP_REPLACE :: StencilOp
STENCIL_OP_REPLACE sets the value to reference.

pattern STENCIL_OP_INCREMENT_AND_CLAMP :: StencilOp
STENCIL_OP_INCREMENT_AND_CLAMP increments the current value and clamps to the maximum representable unsigned value.

pattern STENCIL_OP_DECREMENT_AND_CLAMP :: StencilOp
STENCIL_OP_DECREMENT_AND_CLAMP decrements the current value and clamps to 0.

pattern STENCIL_OP_INVERT :: StencilOp
STENCIL_OP_INVERT bitwise-inverts the current value.

pattern STENCIL_OP_INCREMENT_AND_WRAP :: StencilOp
STENCIL_OP_INCREMENT_AND_WRAP increments the current value and wraps to 0 when the maximum value would have been exceeded.

pattern STENCIL_OP_DECREMENT_AND_WRAP :: StencilOp
STENCIL_OP_DECREMENT_AND_WRAP decrements the current value and wraps to the maximum possible value when the value would go below 0.