Vulkan.Core10.Enums.SubpassContents

Documentation

newtype SubpassContents

VkSubpassContents - Specify how commands in the first subpass of a render pass are provided

See Also

SubpassBeginInfo, cmdBeginRenderPass, cmdNextSubpass

Constructors

SubpassContents Int32

Bundled Patterns

pattern SUBPASS_CONTENTS_INLINE :: SubpassContents
SUBPASS_CONTENTS_INLINE specifies that the contents of the subpass will be recorded inline in the primary command buffer, and secondary command buffers must not be executed within the subpass.

pattern SUBPASS_CONTENTS_SECONDARY_COMMAND_BUFFERS :: SubpassContents
SUBPASS_CONTENTS_SECONDARY_COMMAND_BUFFERS specifies that the contents are recorded in secondary command buffers that will be called from the primary command buffer, and cmdExecuteCommands is the only valid command on the command buffer until cmdNextSubpass or cmdEndRenderPass.