Vulkan.Core10.Enums.VendorId

Documentation

newtype VendorId

VkVendorId - Khronos vendor IDs

Description

Note

Khronos vendor IDs may be allocated by vendors at any time. Only the latest canonical versions of this Specification, of the corresponding vk.xml API Registry, and of the corresponding vulkan_core.h header file must contain all reserved Khronos vendor IDs.

Only Khronos vendor IDs are given symbolic names at present. PCI vendor IDs returned by the implementation can be looked up in the PCI-SIG database.

See Also

No cross-references are available

Constructors

VendorId Int32

Bundled Patterns

pattern VENDOR_ID_VIV :: VendorId

pattern VENDOR_ID_VSI :: VendorId

pattern VENDOR_ID_KAZAN :: VendorId

pattern VENDOR_ID_CODEPLAY :: VendorId

pattern VENDOR_ID_MESA :: VendorId