Vulkan.Core10.Event

Documentation

createEvent

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that creates the event.

-> EventCreateInfo
pCreateInfo is a pointer to a EventCreateInfo structure containing information about how the event is to be created.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io Event

vkCreateEvent - Create a new event object

Description

When created, the event object is in the unsignaled state.

Valid Usage (Implicit)

device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid EventCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pEvent must be a valid pointer to a Event handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, Device, Event, EventCreateInfo

withEvent :: forall io r. MonadIO io => Device -> EventCreateInfo -> Maybe AllocationCallbacks -> (io Event -> (Event -> io ()) -> r) -> r

A convenience wrapper to make a compatible pair of calls to createEvent and destroyEvent

To ensure that destroyEvent is always called: pass bracket (or the allocate function from your favourite resource management library) as the first argument. To just extract the pair pass (,) as the first argument.

destroyEvent

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that destroys the event.

-> Event
event is the handle of the event to destroy.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io ()

vkDestroyEvent - Destroy an event object

Valid Usage

All submitted commands that refer to event must have completed execution

If AllocationCallbacks were provided when event was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when event was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid Device handle

If event is not NULL_HANDLE, event must be a valid Event handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If event is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to event must be externally synchronized

See Also

AllocationCallbacks, Device, Event

getEventStatus

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that owns the event.

device must be a valid Device handle

-> Event
event is the handle of the event to query.

event must be a valid Event handle

event must have been created, allocated, or retrieved from device

-> io Result

vkGetEventStatus - Retrieve the status of an event object

Description

Upon success, getEventStatus returns the state of the event object with the following return codes:

Status 	 Meaning

EVENT_SET 	 The event specified by event is signaled.

EVENT_RESET 	 The event specified by event is unsignaled.

Event Object Status Codes

If a cmdSetEvent or cmdResetEvent command is in a command buffer that is in the pending state, then the value returned by this command may immediately be out of date.

The state of an event can be updated by the host. The state of the event is immediately changed, and subsequent calls to getEventStatus will return the new state. If an event is already in the requested state, then updating it to the same state has no effect.

Return Codes

Success

EVENT_SET

EVENT_RESET

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_DEVICE_LOST

See Also

Device, Event

setEvent

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that owns the event.

-> Event
event is the event to set.

-> io ()

vkSetEvent - Set an event to signaled state

Description

When setEvent is executed on the host, it defines an /event signal operation/ which sets the event to the signaled state.

If event is already in the signaled state when setEvent is executed, then setEvent has no effect, and no event signal operation occurs.

Valid Usage (Implicit)

device must be a valid Device handle

event must be a valid Event handle

event must have been created, allocated, or retrieved from device

Host Synchronization

Host access to event must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

Device, Event

resetEvent

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that owns the event.

-> Event
event is the event to reset.

-> io ()

vkResetEvent - Reset an event to non-signaled state

Description

When resetEvent is executed on the host, it defines an /event unsignal operation/ which resets the event to the unsignaled state.

If event is already in the unsignaled state when resetEvent is executed, then resetEvent has no effect, and no event unsignal operation occurs.

Valid Usage

event must not be waited on by a cmdWaitEvents command that is currently executing

Valid Usage (Implicit)

device must be a valid Device handle

event must be a valid Event handle

event must have been created, allocated, or retrieved from device

Host Synchronization

Host access to event must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

Device, Event

data EventCreateInfo

VkEventCreateInfo - Structure specifying parameters of a newly created event

Valid Usage (Implicit)

See Also

EventCreateFlags, StructureType, createEvent

Constructors

EventCreateInfo

Fields

flags :: EventCreateFlags

flags is reserved for future use.

flags must be 0

newtype Event

VkEvent - Opaque handle to an event object

See Also

cmdResetEvent, cmdSetEvent, cmdWaitEvents, createEvent, destroyEvent, getEventStatus, resetEvent, setEvent

Constructors

Event Word64

newtype EventCreateFlags

VkEventCreateFlags - Reserved for future use

Description

EventCreateFlags is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

EventCreateInfo

Constructors

EventCreateFlags Flags