Vulkan.Core10.ExtensionDiscovery

Documentation

enumerateInstanceExtensionProperties

Arguments

:: forall io. MonadIO io

=> ("layerName" ::: Maybe ByteString)
pLayerName is either NULL or a pointer to a null-terminated UTF-8 string naming the layer to retrieve extensions from.

-> io (Result, "properties" ::: Vector ExtensionProperties)

vkEnumerateInstanceExtensionProperties - Returns up to requested number of global extension properties

Description

When pLayerName parameter is NULL, only extensions provided by the Vulkan implementation or by implicitly enabled layers are returned. When pLayerName is the name of a layer, the instance extensions provided by that layer are returned.

If pProperties is NULL, then the number of extensions properties available is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If pPropertyCount is less than the number of extension properties available, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of extensions available, INCOMPLETE will be returned instead of SUCCESS, to indicate that not all the available properties were returned.

Because the list of available layers may change externally between calls to enumerateInstanceExtensionProperties, two calls may retrieve different results if a pLayerName is available in one call but not in another. The extensions supported by a layer may also change between two calls, e.g. if the layer implementation is replaced by a different version between those calls.

Implementations must not advertise any pair of extensions that cannot be enabled together due to behavioral differences, or any extension that cannot be enabled against the advertised version.

Valid Usage (Implicit)

If pLayerName is not NULL, pLayerName must be a null-terminated UTF-8 string

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount ExtensionProperties structures

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_LAYER_NOT_PRESENT

See Also

ExtensionProperties

enumerateDeviceExtensionProperties

Arguments

:: forall io. MonadIO io

=> PhysicalDevice
physicalDevice is the physical device that will be queried.

-> ("layerName" ::: Maybe ByteString)
pLayerName is either NULL or a pointer to a null-terminated UTF-8 string naming the layer to retrieve extensions from.

-> io (Result, "properties" ::: Vector ExtensionProperties)

vkEnumerateDeviceExtensionProperties - Returns properties of available physical device extensions

Description

When pLayerName parameter is NULL, only extensions provided by the Vulkan implementation or by implicitly enabled layers are returned. When pLayerName is the name of a layer, the device extensions provided by that layer are returned.

Implementations must not advertise any pair of extensions that cannot be enabled together due to behavioral differences, or any extension that cannot be enabled against the advertised version.

Valid Usage (Implicit)

physicalDevice must be a valid PhysicalDevice handle

If pLayerName is not NULL, pLayerName must be a null-terminated UTF-8 string

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount ExtensionProperties structures

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_LAYER_NOT_PRESENT

See Also

ExtensionProperties, PhysicalDevice

data ExtensionProperties

VkExtensionProperties - Structure specifying an extension properties

See Also

enumerateDeviceExtensionProperties, enumerateInstanceExtensionProperties

Constructors

ExtensionProperties

Fields

extensionName :: ByteString

extensionName is an array of MAX_EXTENSION_NAME_SIZE char containing a null-terminated UTF-8 string which is the name of the extension.

specVersion :: Word32

specVersion is the version of this extension. It is an integer, incremented with backward compatible changes.