Vulkan.Core10.LayerDiscovery

Documentation

enumerateInstanceLayerProperties :: forall io. MonadIO io => io (Result, "properties" ::: Vector LayerProperties)

vkEnumerateInstanceLayerProperties - Returns up to requested number of global layer properties

Description

If pProperties is NULL, then the number of layer properties available is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If pPropertyCount is less than the number of layer properties available, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of layers available, INCOMPLETE will be returned instead of SUCCESS, to indicate that not all the available layer properties were returned.

The list of available layers may change at any time due to actions outside of the Vulkan implementation, so two calls to enumerateInstanceLayerProperties with the same parameters may return different results, or retrieve different pPropertyCount values or pProperties contents. Once an instance has been created, the layers enabled for that instance will continue to be enabled and valid for the lifetime of that instance, even if some of them become unavailable for future instances.

Valid Usage (Implicit)

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount LayerProperties structures

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

LayerProperties

enumerateDeviceLayerProperties :: forall io. MonadIO io => PhysicalDevice -> io (Result, "properties" ::: Vector LayerProperties)

vkEnumerateDeviceLayerProperties - Returns properties of available physical device layers

Description

If pProperties is NULL, then the number of layer properties available is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If pPropertyCount is less than the number of layer properties available, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of layers available, INCOMPLETE will be returned instead of SUCCESS, to indicate that not all the available layer properties were returned.

The list of layers enumerated by enumerateDeviceLayerProperties must be exactly the sequence of layers enabled for the instance. The members of LayerProperties for each enumerated layer must be the same as the properties when the layer was enumerated by enumerateInstanceLayerProperties.

Valid Usage (Implicit)

physicalDevice must be a valid PhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount LayerProperties structures

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

LayerProperties, PhysicalDevice

data LayerProperties

VkLayerProperties - Structure specifying layer properties

See Also

enumerateDeviceLayerProperties, enumerateInstanceLayerProperties

Constructors

LayerProperties

Fields

layerName :: ByteString

layerName is an array of MAX_EXTENSION_NAME_SIZE char containing a null-terminated UTF-8 string which is the name of the layer. Use this name in the ppEnabledLayerNames array passed in the InstanceCreateInfo structure to enable this layer for an instance.

specVersion :: Word32

specVersion is the Vulkan version the layer was written to, encoded as described in https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#extendingvulkan-coreversions-versionnumbers.

implementationVersion :: Word32

implementationVersion is the version of this layer. It is an integer, increasing with backward compatible changes.

description :: ByteString

description is an array of MAX_DESCRIPTION_SIZE char containing a null-terminated UTF-8 string which provides additional details that can be used by the application to identify the layer.