Vulkan.Core10.MemoryManagement

Documentation

getBufferMemoryRequirements

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that owns the buffer.

device must be a valid Device handle

-> Buffer
buffer is the buffer to query.

buffer must be a valid Buffer handle

buffer must have been created, allocated, or retrieved from device

-> io MemoryRequirements

vkGetBufferMemoryRequirements - Returns the memory requirements for specified Vulkan object

Valid Usage (Implicit)

See Also

Buffer, Device, MemoryRequirements

bindBufferMemory

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that owns the buffer and memory.

-> Buffer
buffer is the buffer to be attached to memory.

-> DeviceMemory
memory is a DeviceMemory object describing the device memory to attach.

-> ("memoryOffset" ::: DeviceSize)
memoryOffset is the start offset of the region of memory which is to be bound to the buffer. The number of bytes returned in the MemoryRequirements::size member in memory, starting from memoryOffset bytes, will be bound to the specified buffer.

-> io ()

vkBindBufferMemory - Bind device memory to a buffer object

Description

bindBufferMemory is equivalent to passing the same parameters through BindBufferMemoryInfo to bindBufferMemory2.

Valid Usage

buffer must not already be backed by a memory object

buffer must not have been created with any sparse memory binding flags

memoryOffset must be less than the size of memory

memory must have been allocated using one of the memory types allowed in the memoryTypeBits member of the MemoryRequirements structure returned from a call to getBufferMemoryRequirements with buffer

memoryOffset must be an integer multiple of the alignment member of the MemoryRequirements structure returned from a call to getBufferMemoryRequirements with buffer

The size member of the MemoryRequirements structure returned from a call to getBufferMemoryRequirements with buffer must be less than or equal to the size of memory minus memoryOffset

If buffer requires a dedicated allocation(as reported by getBufferMemoryRequirements2 in MemoryDedicatedRequirements::requiresDedicatedAllocation for buffer), memory must have been created with MemoryDedicatedAllocateInfo::buffer equal to buffer

If the MemoryAllocateInfo provided when memory was allocated included a MemoryDedicatedAllocateInfo structure in its pNext chain, and MemoryDedicatedAllocateInfo::buffer was not NULL_HANDLE, then buffer must equal MemoryDedicatedAllocateInfo::buffer, and memoryOffset must be zero

If buffer was created with the BUFFER_CREATE_PROTECTED_BIT bit set, the buffer must be bound to a memory object allocated with a memory type that reports MEMORY_PROPERTY_PROTECTED_BIT

If buffer was created with the BUFFER_CREATE_PROTECTED_BIT bit not set, the buffer must not be bound to a memory object created with a memory type that reports MEMORY_PROPERTY_PROTECTED_BIT

If buffer was created with DedicatedAllocationBufferCreateInfoNV::dedicatedAllocation equal to TRUE, memory must have been created with DedicatedAllocationMemoryAllocateInfoNV::buffer equal to a buffer handle created with identical creation parameters to buffer and memoryOffset must be zero

If the value of ExportMemoryAllocateInfo::handleTypes used to allocate memory is not 0, it must include at least one of the handles set in ExternalMemoryBufferCreateInfo::handleTypes when buffer was created

If memory was created by a memory import operation, that is not ImportAndroidHardwareBufferInfoANDROID with a non-NULL buffer value, the external handle type of the imported memory must also have been set in ExternalMemoryBufferCreateInfo::handleTypes when buffer was created

If memory was created with the ImportAndroidHardwareBufferInfoANDROID memory import operation with a non-NULL buffer value, EXTERNAL_MEMORY_HANDLE_TYPE_ANDROID_HARDWARE_BUFFER_BIT_ANDROID must also have been set in ExternalMemoryBufferCreateInfo::handleTypes when buffer was created

If the PhysicalDeviceBufferDeviceAddressFeatures::bufferDeviceAddress feature is enabled and buffer was created with the BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT bit set, memory must have been allocated with the MEMORY_ALLOCATE_DEVICE_ADDRESS_BIT bit set

Valid Usage (Implicit)

device must be a valid Device handle

buffer must be a valid Buffer handle

memory must be a valid DeviceMemory handle

buffer must have been created, allocated, or retrieved from device

memory must have been created, allocated, or retrieved from device

Host Synchronization

Host access to buffer must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_INVALID_OPAQUE_CAPTURE_ADDRESS_KHR

See Also

Buffer, Device, DeviceMemory, DeviceSize

getImageMemoryRequirements

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that owns the image.

-> Image
image is the image to query.

-> io MemoryRequirements

vkGetImageMemoryRequirements - Returns the memory requirements for specified Vulkan object

Valid Usage

image must not have been created with the IMAGE_CREATE_DISJOINT_BIT flag set

If image was created with the EXTERNAL_MEMORY_HANDLE_TYPE_ANDROID_HARDWARE_BUFFER_BIT_ANDROID external memory handle type, then image must be bound to memory

Valid Usage (Implicit)

device must be a valid Device handle

image must be a valid Image handle

pMemoryRequirements must be a valid pointer to a MemoryRequirements structure

image must have been created, allocated, or retrieved from device

See Also

Device, Image, MemoryRequirements

bindImageMemory

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that owns the image and memory.

-> Image
image is the image.

-> DeviceMemory
memory is the DeviceMemory object describing the device memory to attach.

-> ("memoryOffset" ::: DeviceSize)
memoryOffset is the start offset of the region of memory which is to be bound to the image. The number of bytes returned in the MemoryRequirements::size member in memory, starting from memoryOffset bytes, will be bound to the specified image.

-> io ()

vkBindImageMemory - Bind device memory to an image object

Description

bindImageMemory is equivalent to passing the same parameters through BindImageMemoryInfo to bindImageMemory2.

Valid Usage

image must not have been created with the IMAGE_CREATE_DISJOINT_BIT set

image must not already be backed by a memory object

image must not have been created with any sparse memory binding flags

memoryOffset must be less than the size of memory

memory must have been allocated using one of the memory types allowed in the memoryTypeBits member of the MemoryRequirements structure returned from a call to getImageMemoryRequirements with image

memoryOffset must be an integer multiple of the alignment member of the MemoryRequirements structure returned from a call to getImageMemoryRequirements with image

The difference of the size of memory and memoryOffset must be greater than or equal to the size member of the MemoryRequirements structure returned from a call to getImageMemoryRequirements with the same image

If image requires a dedicated allocation (as reported by getImageMemoryRequirements2 in MemoryDedicatedRequirements::requiresDedicatedAllocation for image), memory must have been created with MemoryDedicatedAllocateInfo::image equal to image

If the dedicated allocation image aliasing feature is not enabled, and the MemoryAllocateInfo provided when memory was allocated included a MemoryDedicatedAllocateInfo structure in its pNext chain, and MemoryDedicatedAllocateInfo::image was not NULL_HANDLE, then image must equal MemoryDedicatedAllocateInfo::image and memoryOffset must be zero

If the dedicated allocation image aliasing feature is enabled, and the MemoryAllocateInfo provided when memory was allocated included a MemoryDedicatedAllocateInfo structure in its pNext chain, and MemoryDedicatedAllocateInfo::image was not NULL_HANDLE, then memoryOffset must be zero, and image must be either equal to MemoryDedicatedAllocateInfo::image or an image that was created using the same parameters in ImageCreateInfo, with the exception that extent and arrayLayers may differ subject to the following restrictions: every dimension in the extent parameter of the image being bound must be equal to or smaller than the original image for which the allocation was created; and the arrayLayers parameter of the image being bound must be equal to or smaller than the original image for which the allocation was created

If image was created with the IMAGE_CREATE_PROTECTED_BIT bit set, the image must be bound to a memory object allocated with a memory type that reports MEMORY_PROPERTY_PROTECTED_BIT

If image was created with the IMAGE_CREATE_PROTECTED_BIT bit not set, the image must not be bound to a memory object created with a memory type that reports MEMORY_PROPERTY_PROTECTED_BIT

If image was created with DedicatedAllocationImageCreateInfoNV::dedicatedAllocation equal to TRUE, memory must have been created with DedicatedAllocationMemoryAllocateInfoNV::image equal to an image handle created with identical creation parameters to image and memoryOffset must be zero

If the value of ExportMemoryAllocateInfo::handleTypes used to allocate memory is not 0, it must include at least one of the handles set in ExternalMemoryImageCreateInfo::handleTypes when image was created

If memory was created by a memory import operation, that is not ImportAndroidHardwareBufferInfoANDROID with a non-NULL buffer value, the external handle type of the imported memory must also have been set in ExternalMemoryImageCreateInfo::handleTypes when image was created

If memory was created with the ImportAndroidHardwareBufferInfoANDROID memory import operation with a non-NULL buffer value, EXTERNAL_MEMORY_HANDLE_TYPE_ANDROID_HARDWARE_BUFFER_BIT_ANDROID must also have been set in ExternalMemoryImageCreateInfo::handleTypes when image was created

Valid Usage (Implicit)

device must be a valid Device handle

image must be a valid Image handle

memory must be a valid DeviceMemory handle

image must have been created, allocated, or retrieved from device

memory must have been created, allocated, or retrieved from device

Host Synchronization

Host access to image must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

Device, DeviceMemory, DeviceSize, Image

data MemoryRequirements

VkMemoryRequirements - Structure specifying memory requirements

See Also

DeviceSize, MemoryRequirements2, getBufferMemoryRequirements, getImageMemoryRequirements

Constructors

MemoryRequirements

Fields

size :: DeviceSize

size is the size, in bytes, of the memory allocation required for the resource.

alignment :: DeviceSize

alignment is the alignment, in bytes, of the offset within the allocation required for the resource.

memoryTypeBits :: Word32

memoryTypeBits is a bitmask and contains one bit set for every supported memory type for the resource. Bit i is set if and only if the memory type i in the PhysicalDeviceMemoryProperties structure for the physical device is supported for the resource.

newtype DeviceMemory

VkDeviceMemory - Opaque handle to a device memory object

See Also

BindAccelerationStructureMemoryInfoKHR, BindBufferMemoryInfo, BindImageMemoryInfo, DeviceMemoryOpaqueCaptureAddressInfo, MappedMemoryRange, MemoryGetAndroidHardwareBufferInfoANDROID, MemoryGetFdInfoKHR, MemoryGetWin32HandleInfoKHR, SparseImageMemoryBind, SparseMemoryBind, Win32KeyedMutexAcquireReleaseInfoKHR, Win32KeyedMutexAcquireReleaseInfoNV, allocateMemory, bindBufferMemory, bindImageMemory, freeMemory, getDeviceMemoryCommitment, getMemoryWin32HandleNV, mapMemory, unmapMemory

Constructors

DeviceMemory Word64