Vulkan.Core10.PipelineLayout

Documentation

createPipelineLayout

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that creates the pipeline layout.

-> PipelineLayoutCreateInfo
pCreateInfo is a pointer to a PipelineLayoutCreateInfo structure specifying the state of the pipeline layout object.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io PipelineLayout

vkCreatePipelineLayout - Creates a new pipeline layout object

Valid Usage (Implicit)

device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid PipelineLayoutCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pPipelineLayout must be a valid pointer to a PipelineLayout handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, Device, PipelineLayout, PipelineLayoutCreateInfo

withPipelineLayout :: forall io r. MonadIO io => Device -> PipelineLayoutCreateInfo -> Maybe AllocationCallbacks -> (io PipelineLayout -> (PipelineLayout -> io ()) -> r) -> r

A convenience wrapper to make a compatible pair of calls to createPipelineLayout and destroyPipelineLayout

To ensure that destroyPipelineLayout is always called: pass bracket (or the allocate function from your favourite resource management library) as the first argument. To just extract the pair pass (,) as the first argument.

destroyPipelineLayout

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that destroys the pipeline layout.

-> PipelineLayout
pipelineLayout is the pipeline layout to destroy.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io ()

vkDestroyPipelineLayout - Destroy a pipeline layout object

Valid Usage

If AllocationCallbacks were provided when pipelineLayout was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when pipelineLayout was created, pAllocator must be NULL

pipelineLayout must not have been passed to any vkCmd* command for any command buffers that are still in the recording state when destroyPipelineLayout is called

Valid Usage (Implicit)

device must be a valid Device handle

If pipelineLayout is not NULL_HANDLE, pipelineLayout must be a valid PipelineLayout handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If pipelineLayout is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to pipelineLayout must be externally synchronized

See Also

AllocationCallbacks, Device, PipelineLayout

data PushConstantRange

VkPushConstantRange - Structure specifying a push constant range

Valid Usage (Implicit)

See Also

PipelineLayoutCreateInfo, ShaderStageFlags

Constructors

PushConstantRange

Fields

stageFlags :: ShaderStageFlags

stageFlags is a set of stage flags describing the shader stages that will access a range of push constants. If a particular stage is not included in the range, then accessing members of that range of push constants from the corresponding shader stage will return undefined values.

stageFlags must be a valid combination of ShaderStageFlagBits values

stageFlags must not be 0

offset :: Word32

offset and size are the start offset and size, respectively, consumed by the range. Both offset and size are in units of bytes and must be a multiple of 4. The layout of the push constant variables is specified in the shader.

offset must be less than PhysicalDeviceLimits::maxPushConstantsSize

offset must be a multiple of 4

size :: Word32

size must be greater than 0

size must be a multiple of 4

size must be less than or equal to PhysicalDeviceLimits::maxPushConstantsSize minus offset

data PipelineLayoutCreateInfo

VkPipelineLayoutCreateInfo - Structure specifying the parameters of a newly created pipeline layout object

Valid Usage

setLayoutCount must be less than or equal to PhysicalDeviceLimits::maxBoundDescriptorSets

The total number of descriptors in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set with a descriptorType of DESCRIPTOR_TYPE_SAMPLER and DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER accessible to any given shader stage across all elements of pSetLayouts must be less than or equal to PhysicalDeviceLimits::maxPerStageDescriptorSamplers

The total number of descriptors in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set with a descriptorType of DESCRIPTOR_TYPE_UNIFORM_BUFFER and DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC accessible to any given shader stage across all elements of pSetLayouts must be less than or equal to PhysicalDeviceLimits::maxPerStageDescriptorUniformBuffers

The total number of descriptors in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set with a descriptorType of DESCRIPTOR_TYPE_STORAGE_BUFFER and DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC accessible to any given shader stage across all elements of pSetLayouts must be less than or equal to PhysicalDeviceLimits::maxPerStageDescriptorStorageBuffers

The total number of descriptors in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set with a descriptorType of DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, DESCRIPTOR_TYPE_SAMPLED_IMAGE, and DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER accessible to any given shader stage across all elements of pSetLayouts must be less than or equal to PhysicalDeviceLimits::maxPerStageDescriptorSampledImages

The total number of descriptors in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set with a descriptorType of DESCRIPTOR_TYPE_STORAGE_IMAGE, and DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER accessible to any given shader stage across all elements of pSetLayouts must be less than or equal to PhysicalDeviceLimits::maxPerStageDescriptorStorageImages

The total number of descriptors in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set with a descriptorType of DESCRIPTOR_TYPE_INPUT_ATTACHMENT accessible to any given shader stage across all elements of pSetLayouts must be less than or equal to PhysicalDeviceLimits::maxPerStageDescriptorInputAttachments

The total number of bindings in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set with a descriptorType of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT accessible to any given shader stage across all elements of pSetLayouts must be less than or equal to PhysicalDeviceInlineUniformBlockPropertiesEXT::maxPerStageDescriptorInlineUniformBlocks

The total number of descriptors with a descriptorType of DESCRIPTOR_TYPE_SAMPLER and DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER accessible to any given shader stage across all elements of pSetLayouts must be less than or equal to PhysicalDeviceDescriptorIndexingProperties::maxPerStageDescriptorUpdateAfterBindSamplers

The total number of descriptors with a descriptorType of DESCRIPTOR_TYPE_UNIFORM_BUFFER and DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC accessible to any given shader stage across all elements of pSetLayouts must be less than or equal to PhysicalDeviceDescriptorIndexingProperties::maxPerStageDescriptorUpdateAfterBindUniformBuffers

The total number of descriptors with a descriptorType of DESCRIPTOR_TYPE_STORAGE_BUFFER and DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC accessible to any given shader stage across all elements of pSetLayouts must be less than or equal to PhysicalDeviceDescriptorIndexingProperties::maxPerStageDescriptorUpdateAfterBindStorageBuffers

The total number of descriptors with a descriptorType of DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, DESCRIPTOR_TYPE_SAMPLED_IMAGE, and DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER accessible to any given shader stage across all elements of pSetLayouts must be less than or equal to PhysicalDeviceDescriptorIndexingProperties::maxPerStageDescriptorUpdateAfterBindSampledImages

The total number of descriptors with a descriptorType of DESCRIPTOR_TYPE_STORAGE_IMAGE, and DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER accessible to any given shader stage across all elements of pSetLayouts must be less than or equal to PhysicalDeviceDescriptorIndexingProperties::maxPerStageDescriptorUpdateAfterBindStorageImages

The total number of descriptors with a descriptorType of DESCRIPTOR_TYPE_INPUT_ATTACHMENT accessible to any given shader stage across all elements of pSetLayouts must be less than or equal to PhysicalDeviceDescriptorIndexingProperties::maxPerStageDescriptorUpdateAfterBindInputAttachments

The total number of bindings with a descriptorType of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT accessible to any given shader stage across all elements of pSetLayouts must be less than or equal to PhysicalDeviceInlineUniformBlockPropertiesEXT::maxPerStageDescriptorUpdateAfterBindInlineUniformBlocks

The total number of descriptors in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set with a descriptorType of DESCRIPTOR_TYPE_SAMPLER and DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceLimits::maxDescriptorSetSamplers

The total number of descriptors in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set with a descriptorType of DESCRIPTOR_TYPE_UNIFORM_BUFFER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceLimits::maxDescriptorSetUniformBuffers

The total number of descriptors in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set with a descriptorType of DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceLimits::maxDescriptorSetUniformBuffersDynamic

The total number of descriptors in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set with a descriptorType of DESCRIPTOR_TYPE_STORAGE_BUFFER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceLimits::maxDescriptorSetStorageBuffers

The total number of descriptors in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set with a descriptorType of DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceLimits::maxDescriptorSetStorageBuffersDynamic

The total number of descriptors in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set with a descriptorType of DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, DESCRIPTOR_TYPE_SAMPLED_IMAGE, and DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceLimits::maxDescriptorSetSampledImages

The total number of descriptors in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set with a descriptorType of DESCRIPTOR_TYPE_STORAGE_IMAGE, and DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceLimits::maxDescriptorSetStorageImages

The total number of descriptors in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set with a descriptorType of DESCRIPTOR_TYPE_INPUT_ATTACHMENT accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceLimits::maxDescriptorSetInputAttachments

The total number of bindings in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set with a descriptorType of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceInlineUniformBlockPropertiesEXT::maxDescriptorSetInlineUniformBlocks

The total number of descriptors of the type DESCRIPTOR_TYPE_SAMPLER and DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceDescriptorIndexingProperties::maxDescriptorSetUpdateAfterBindSamplers

The total number of descriptors of the type DESCRIPTOR_TYPE_UNIFORM_BUFFER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceDescriptorIndexingProperties::maxDescriptorSetUpdateAfterBindUniformBuffers

The total number of descriptors of the type DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceDescriptorIndexingProperties::maxDescriptorSetUpdateAfterBindUniformBuffersDynamic

The total number of descriptors of the type DESCRIPTOR_TYPE_STORAGE_BUFFER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceDescriptorIndexingProperties::maxDescriptorSetUpdateAfterBindStorageBuffers

The total number of descriptors of the type DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceDescriptorIndexingProperties::maxDescriptorSetUpdateAfterBindStorageBuffersDynamic

The total number of descriptors of the type DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, DESCRIPTOR_TYPE_SAMPLED_IMAGE, and DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceDescriptorIndexingProperties::maxDescriptorSetUpdateAfterBindSampledImages

The total number of descriptors of the type DESCRIPTOR_TYPE_STORAGE_IMAGE, and DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceDescriptorIndexingProperties::maxDescriptorSetUpdateAfterBindStorageImages

The total number of descriptors of the type DESCRIPTOR_TYPE_INPUT_ATTACHMENT accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceDescriptorIndexingProperties::maxDescriptorSetUpdateAfterBindInputAttachments

The total number of bindings with a descriptorType of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceInlineUniformBlockPropertiesEXT::maxDescriptorSetUpdateAfterBindInlineUniformBlocks

Any two elements of pPushConstantRanges must not include the same stage in stageFlags

pSetLayouts must not contain more than one descriptor set layout that was created with DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR set

The total number of bindings with a descriptorType of DESCRIPTOR_TYPE_ACCELERATION_STRUCTURE_KHR accessible across all shader stages and across all elements of pSetLayouts must be less than or equal to PhysicalDeviceRayTracingPropertiesKHR::maxDescriptorSetAccelerationStructures

The total number of pImmutableSamplers created with flags containing SAMPLER_CREATE_SUBSAMPLED_BIT_EXT or SAMPLER_CREATE_SUBSAMPLED_COARSE_RECONSTRUCTION_BIT_EXT across all shader stages and across all elements of pSetLayouts must be less than or equal to ::maxDescriptorSetSubsampledSamplers

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO

pNext must be NULL

flags must be 0

If setLayoutCount is not 0, pSetLayouts must be a valid pointer to an array of setLayoutCount valid DescriptorSetLayout handles

If pushConstantRangeCount is not 0, pPushConstantRanges must be a valid pointer to an array of pushConstantRangeCount valid PushConstantRange structures

See Also

DescriptorSetLayout, PipelineLayoutCreateFlags, PushConstantRange, StructureType, createPipelineLayout

Constructors

PipelineLayoutCreateInfo

Fields

flags :: PipelineLayoutCreateFlags

flags is reserved for future use.

setLayouts :: Vector DescriptorSetLayout

pSetLayouts is a pointer to an array of DescriptorSetLayout objects.

pushConstantRanges :: Vector PushConstantRange

pPushConstantRanges is a pointer to an array of PushConstantRange structures defining a set of push constant ranges for use in a single pipeline layout. In addition to descriptor set layouts, a pipeline layout also describes how many push constants can be accessed by each stage of the pipeline.

Note

Push constants represent a high speed path to modify constant data in pipelines that is expected to outperform memory-backed resource updates.

newtype PipelineLayout

VkPipelineLayout - Opaque handle to a pipeline layout object

See Also

ComputePipelineCreateInfo, DescriptorUpdateTemplateCreateInfo, GraphicsPipelineCreateInfo, IndirectCommandsLayoutTokenNV, RayTracingPipelineCreateInfoKHR, RayTracingPipelineCreateInfoNV, cmdBindDescriptorSets, cmdPushConstants, cmdPushDescriptorSetKHR, cmdPushDescriptorSetWithTemplateKHR, createPipelineLayout, destroyPipelineLayout

Constructors

PipelineLayout Word64