Vulkan.Core10.Query

Documentation

createQueryPool

Arguments

:: forall a io. (Extendss QueryPoolCreateInfo a, PokeChain a, MonadIO io)

=> Device
device is the logical device that creates the query pool.

-> QueryPoolCreateInfo a
pCreateInfo is a pointer to a QueryPoolCreateInfo structure containing the number and type of queries to be managed by the pool.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io QueryPool

vkCreateQueryPool - Create a new query pool object

Valid Usage (Implicit)

device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid QueryPoolCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pQueryPool must be a valid pointer to a QueryPool handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, Device, QueryPool, QueryPoolCreateInfo

withQueryPool :: forall a io r. (Extendss QueryPoolCreateInfo a, PokeChain a, MonadIO io) => Device -> QueryPoolCreateInfo a -> Maybe AllocationCallbacks -> (io QueryPool -> (QueryPool -> io ()) -> r) -> r

A convenience wrapper to make a compatible pair of calls to createQueryPool and destroyQueryPool

To ensure that destroyQueryPool is always called: pass bracket (or the allocate function from your favourite resource management library) as the first argument. To just extract the pair pass (,) as the first argument.

destroyQueryPool

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that destroys the query pool.

-> QueryPool
queryPool is the query pool to destroy.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io ()

vkDestroyQueryPool - Destroy a query pool object

Valid Usage

All submitted commands that refer to queryPool must have completed execution

If AllocationCallbacks were provided when queryPool was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when queryPool was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid Device handle

If queryPool is not NULL_HANDLE, queryPool must be a valid QueryPool handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If queryPool is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to queryPool must be externally synchronized

See Also

AllocationCallbacks, Device, QueryPool

getQueryPoolResults

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that owns the query pool.

-> QueryPool
queryPool is the query pool managing the queries containing the desired results.

-> ("firstQuery" ::: Word32)
firstQuery is the initial query index.

-> ("queryCount" ::: Word32)
queryCount is the number of queries to read.

-> ("dataSize" ::: Word64)
dataSize is the size in bytes of the buffer pointed to by pData.

-> ("data" ::: Ptr ())
pData is a pointer to a user-allocated buffer where the results will be written

-> ("stride" ::: DeviceSize)
stride is the stride in bytes between results for individual queries within pData.

-> QueryResultFlags
flags is a bitmask of QueryResultFlagBits specifying how and when results are returned.

-> io Result

vkGetQueryPoolResults - Copy results of queries in a query pool to a host memory region

Description

The range of queries read is defined by [firstQuery, firstQuery + queryCount - 1]. For pipeline statistics queries, each query index in the pool contains one integer value for each bit that is enabled in QueryPoolCreateInfo::pipelineStatistics when the pool is created.

If no bits are set in flags, and all requested queries are in the available state, results are written as an array of 32-bit unsigned integer values. The behavior when not all queries are available, is described below.

If QUERY_RESULT_64_BIT is not set and the result overflows a 32-bit value, the value may either wrap or saturate. Similarly, if QUERY_RESULT_64_BIT is set and the result overflows a 64-bit value, the value may either wrap or saturate.

If QUERY_RESULT_WAIT_BIT is set, Vulkan will wait for each query to be in the available state before retrieving the numerical results for that query. In this case, getQueryPoolResults is guaranteed to succeed and return SUCCESS if the queries become available in a finite time (i.e. if they have been issued and not reset). If queries will never finish (e.g. due to being reset but not issued), then getQueryPoolResults may not return in finite time.

If QUERY_RESULT_WAIT_BIT and QUERY_RESULT_PARTIAL_BIT are both not set then no result values are written to pData for queries that are in the unavailable state at the time of the call, and getQueryPoolResults returns NOT_READY. However, availability state is still written to pData for those queries if QUERY_RESULT_WITH_AVAILABILITY_BIT is set.

Note

Applications must take care to ensure that use of the QUERY_RESULT_WAIT_BIT bit has the desired effect.

For example, if a query has been used previously and a command buffer records the commands cmdResetQueryPool, cmdBeginQuery, and cmdEndQuery for that query, then the query will remain in the available state until resetQueryPool is called or the cmdResetQueryPool command executes on a queue. Applications can use fences or events to ensure that a query has already been reset before checking for its results or availability status. Otherwise, a stale value could be returned from a previous use of the query.

The above also applies when QUERY_RESULT_WAIT_BIT is used in combination with QUERY_RESULT_WITH_AVAILABILITY_BIT. In this case, the returned availability status may reflect the result of a previous use of the query unless resetQueryPool is called or the cmdResetQueryPool command has been executed since the last use of the query.

Note

Applications can double-buffer query pool usage, with a pool per frame, and reset queries at the end of the frame in which they are read.

If QUERY_RESULT_PARTIAL_BIT is set, QUERY_RESULT_WAIT_BIT is not set, and the query’s status is unavailable, an intermediate result value between zero and the final result value is written to pData for that query.

If QUERY_RESULT_WITH_AVAILABILITY_BIT is set, the final integer value written for each query is non-zero if the query’s status was available or zero if the status was unavailable. When QUERY_RESULT_WITH_AVAILABILITY_BIT is used, implementations must guarantee that if they return a non-zero availability value then the numerical results must be valid, assuming the results are not reset by a subsequent command.

Note

Satisfying this guarantee may require careful ordering by the application, e.g. to read the availability status before reading the results.

Valid Usage

firstQuery must be less than the number of queries in queryPool

If QUERY_RESULT_64_BIT is not set in flags, then pData and stride must be multiples of 4

If QUERY_RESULT_64_BIT is not set in flags and the queryType used to create queryPool was not QUERY_TYPE_PERFORMANCE_QUERY_KHR, then pData and stride must be multiples of 4

If QUERY_RESULT_64_BIT is set in flags then pData and stride must be multiples of 8

If the queryType used to create queryPool was QUERY_TYPE_PERFORMANCE_QUERY_KHR, then pData and stride must be multiples of the size of PerformanceCounterResultKHR

The sum of firstQuery and queryCount must be less than or equal to the number of queries in queryPool

dataSize must be large enough to contain the result of each query, as described here

If the queryType used to create queryPool was QUERY_TYPE_TIMESTAMP, flags must not contain QUERY_RESULT_PARTIAL_BIT

If the queryType used to create queryPool was QUERY_TYPE_PERFORMANCE_QUERY_KHR, flags must not contain QUERY_RESULT_WITH_AVAILABILITY_BIT, QUERY_RESULT_PARTIAL_BIT or QUERY_RESULT_64_BIT

If the queryType used to create queryPool was QUERY_TYPE_PERFORMANCE_QUERY_KHR, the queryPool must have been recorded once for each pass as retrieved via a call to getPhysicalDeviceQueueFamilyPerformanceQueryPassesKHR

Valid Usage (Implicit)

device must be a valid Device handle

queryPool must be a valid QueryPool handle

pData must be a valid pointer to an array of dataSize bytes

flags must be a valid combination of QueryResultFlagBits values

dataSize must be greater than 0

queryPool must have been created, allocated, or retrieved from device

Return Codes

Success

SUCCESS

NOT_READY

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_DEVICE_LOST

See Also

Device, DeviceSize, QueryPool, QueryResultFlags

data QueryPoolCreateInfo (es :: [Type])

VkQueryPoolCreateInfo - Structure specifying parameters of a newly created query pool

Description

pipelineStatistics is ignored if queryType is not QUERY_TYPE_PIPELINE_STATISTICS.

Valid Usage

If the pipeline statistics queries feature is not enabled, queryType must not be QUERY_TYPE_PIPELINE_STATISTICS

If queryType is QUERY_TYPE_PIPELINE_STATISTICS, pipelineStatistics must be a valid combination of QueryPipelineStatisticFlagBits values

If queryType is QUERY_TYPE_PERFORMANCE_QUERY_KHR, the pNext chain must include a structure of type QueryPoolPerformanceCreateInfoKHR

queryCount must be greater than 0

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_QUERY_POOL_CREATE_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of QueryPoolPerformanceCreateInfoKHR or QueryPoolPerformanceQueryCreateInfoINTEL

The sType value of each struct in the pNext chain must be unique

flags must be 0

queryType must be a valid QueryType value

See Also

QueryPipelineStatisticFlags, QueryPoolCreateFlags, QueryType, StructureType, createQueryPool

Constructors

QueryPoolCreateInfo

Fields

next :: Chain es

pNext is NULL or a pointer to a structure extending this structure.

flags :: QueryPoolCreateFlags

flags is reserved for future use.

queryType :: QueryType

queryType is a QueryType value specifying the type of queries managed by the pool.

queryCount :: Word32

queryCount is the number of queries managed by the pool.

pipelineStatistics :: QueryPipelineStatisticFlags

pipelineStatistics is a bitmask of QueryPipelineStatisticFlagBits specifying which counters will be returned in queries on the new pool, as described below in https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#queries-pipestats.

newtype QueryPool

VkQueryPool - Opaque handle to a query pool object

See Also

cmdBeginQuery, cmdBeginQueryIndexedEXT, cmdCopyQueryPoolResults, cmdEndQuery, cmdEndQueryIndexedEXT, cmdResetQueryPool, cmdWriteAccelerationStructuresPropertiesKHR, cmdWriteAccelerationStructuresPropertiesNV, cmdWriteTimestamp, createQueryPool, destroyQueryPool, getQueryPoolResults, resetQueryPool, resetQueryPoolEXT

Constructors

QueryPool Word64

newtype QueryPoolCreateFlags

VkQueryPoolCreateFlags - Reserved for future use

Description

QueryPoolCreateFlags is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

QueryPoolCreateInfo

Constructors

QueryPoolCreateFlags Flags

newtype QueryType

VkQueryType - Specify the type of queries managed by a query pool

See Also

QueryPoolCreateInfo, cmdWriteAccelerationStructuresPropertiesKHR, cmdWriteAccelerationStructuresPropertiesNV, writeAccelerationStructuresPropertiesKHR

Constructors

QueryType Int32

Bundled Patterns

pattern QUERY_TYPE_OCCLUSION :: QueryType
QUERY_TYPE_OCCLUSION specifies an occlusion query.

pattern QUERY_TYPE_PIPELINE_STATISTICS :: QueryType
QUERY_TYPE_PIPELINE_STATISTICS specifies a pipeline statistics query.

pattern QUERY_TYPE_TIMESTAMP :: QueryType
QUERY_TYPE_TIMESTAMP specifies a timestamp query.

pattern QUERY_TYPE_PERFORMANCE_QUERY_INTEL :: QueryType
QUERY_TYPE_PERFORMANCE_QUERY_INTEL specifies a Intel performance query.

pattern QUERY_TYPE_ACCELERATION_STRUCTURE_SERIALIZATION_SIZE_KHR :: QueryType
QUERY_TYPE_ACCELERATION_STRUCTURE_SERIALIZATION_SIZE_KHR specifies a ray tracing serialization acceleration structure size query

pattern QUERY_TYPE_ACCELERATION_STRUCTURE_COMPACTED_SIZE_KHR :: QueryType
QUERY_TYPE_ACCELERATION_STRUCTURE_COMPACTED_SIZE_KHR specifies a ray tracing acceleration structure size query.

pattern QUERY_TYPE_PERFORMANCE_QUERY_KHR :: QueryType
QUERY_TYPE_PERFORMANCE_QUERY_KHR specifies a performance query.

pattern QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT :: QueryType
QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT specifies a transform feedback query.

newtype QueryResultFlagBits

VkQueryResultFlagBits - Bitmask specifying how and when query results are returned

See Also

QueryResultFlags

Constructors

QueryResultFlagBits Flags

Bundled Patterns

pattern QUERY_RESULT_64_BIT :: QueryResultFlagBits
QUERY_RESULT_64_BIT specifies the results will be written as an array of 64-bit unsigned integer values. If this bit is not set, the results will be written as an array of 32-bit unsigned integer values.

pattern QUERY_RESULT_WAIT_BIT :: QueryResultFlagBits
QUERY_RESULT_WAIT_BIT specifies that Vulkan will wait for each query’s status to become available before retrieving its results.

pattern QUERY_RESULT_WITH_AVAILABILITY_BIT :: QueryResultFlagBits
QUERY_RESULT_WITH_AVAILABILITY_BIT specifies that the availability status accompanies the results.

pattern QUERY_RESULT_PARTIAL_BIT :: QueryResultFlagBits
QUERY_RESULT_PARTIAL_BIT specifies that returning partial results is acceptable.

type QueryResultFlags = QueryResultFlagBits

newtype QueryPipelineStatisticFlagBits

VkQueryPipelineStatisticFlagBits - Bitmask specifying queried pipeline statistics

Description

These values are intended to measure relative statistics on one implementation. Various device architectures will count these values differently. Any or all counters may be affected by the issues described in Query Operation.

Note

For example, tile-based rendering devices may need to replay the scene multiple times, affecting some of the counts.

If a pipeline has rasterizerDiscardEnable enabled, implementations may discard primitives after the final vertex processing stage. As a result, if rasterizerDiscardEnable is enabled, the clipping input and output primitives counters may not be incremented.

When a pipeline statistics query finishes, the result for that query is marked as available. The application can copy the result to a buffer (via cmdCopyQueryPoolResults), or request it be put into host memory (via getQueryPoolResults).

See Also

QueryPipelineStatisticFlags

Constructors

QueryPipelineStatisticFlagBits Flags

Bundled Patterns

pattern QUERY_PIPELINE_STATISTIC_INPUT_ASSEMBLY_VERTICES_BIT :: QueryPipelineStatisticFlagBits
QUERY_PIPELINE_STATISTIC_INPUT_ASSEMBLY_VERTICES_BIT specifies that queries managed by the pool will count the number of vertices processed by the input assembly stage. Vertices corresponding to incomplete primitives may contribute to the count.

pattern QUERY_PIPELINE_STATISTIC_INPUT_ASSEMBLY_PRIMITIVES_BIT :: QueryPipelineStatisticFlagBits
QUERY_PIPELINE_STATISTIC_INPUT_ASSEMBLY_PRIMITIVES_BIT specifies that queries managed by the pool will count the number of primitives processed by the input assembly stage. If primitive restart is enabled, restarting the primitive topology has no effect on the count. Incomplete primitives may be counted.

pattern QUERY_PIPELINE_STATISTIC_VERTEX_SHADER_INVOCATIONS_BIT :: QueryPipelineStatisticFlagBits
QUERY_PIPELINE_STATISTIC_VERTEX_SHADER_INVOCATIONS_BIT specifies that queries managed by the pool will count the number of vertex shader invocations. This counter’s value is incremented each time a vertex shader is invoked.

pattern QUERY_PIPELINE_STATISTIC_GEOMETRY_SHADER_INVOCATIONS_BIT :: QueryPipelineStatisticFlagBits
QUERY_PIPELINE_STATISTIC_GEOMETRY_SHADER_INVOCATIONS_BIT specifies that queries managed by the pool will count the number of geometry shader invocations. This counter’s value is incremented each time a geometry shader is invoked. In the case of instanced geometry shaders, the geometry shader invocations count is incremented for each separate instanced invocation.

pattern QUERY_PIPELINE_STATISTIC_GEOMETRY_SHADER_PRIMITIVES_BIT :: QueryPipelineStatisticFlagBits
QUERY_PIPELINE_STATISTIC_GEOMETRY_SHADER_PRIMITIVES_BIT specifies that queries managed by the pool will count the number of primitives generated by geometry shader invocations. The counter’s value is incremented each time the geometry shader emits a primitive. Restarting primitive topology using the SPIR-V instructions OpEndPrimitive or OpEndStreamPrimitive has no effect on the geometry shader output primitives count.

pattern QUERY_PIPELINE_STATISTIC_CLIPPING_INVOCATIONS_BIT :: QueryPipelineStatisticFlagBits
QUERY_PIPELINE_STATISTIC_CLIPPING_INVOCATIONS_BIT specifies that queries managed by the pool will count the number of primitives processed by the Primitive Clipping stage of the pipeline. The counter’s value is incremented each time a primitive reaches the primitive clipping stage.

pattern QUERY_PIPELINE_STATISTIC_CLIPPING_PRIMITIVES_BIT :: QueryPipelineStatisticFlagBits
QUERY_PIPELINE_STATISTIC_CLIPPING_PRIMITIVES_BIT specifies that queries managed by the pool will count the number of primitives output by the Primitive Clipping stage of the pipeline. The counter’s value is incremented each time a primitive passes the primitive clipping stage. The actual number of primitives output by the primitive clipping stage for a particular input primitive is implementation-dependent but must satisfy the following conditions:

If at least one vertex of the input primitive lies inside the clipping volume, the counter is incremented by one or more.

Otherwise, the counter is incremented by zero or more.

pattern QUERY_PIPELINE_STATISTIC_FRAGMENT_SHADER_INVOCATIONS_BIT :: QueryPipelineStatisticFlagBits
QUERY_PIPELINE_STATISTIC_FRAGMENT_SHADER_INVOCATIONS_BIT specifies that queries managed by the pool will count the number of fragment shader invocations. The counter’s value is incremented each time the fragment shader is invoked.

pattern QUERY_PIPELINE_STATISTIC_TESSELLATION_CONTROL_SHADER_PATCHES_BIT :: QueryPipelineStatisticFlagBits
QUERY_PIPELINE_STATISTIC_TESSELLATION_CONTROL_SHADER_PATCHES_BIT specifies that queries managed by the pool will count the number of patches processed by the tessellation control shader. The counter’s value is incremented once for each patch for which a tessellation control shader is invoked.

pattern QUERY_PIPELINE_STATISTIC_TESSELLATION_EVALUATION_SHADER_INVOCATIONS_BIT :: QueryPipelineStatisticFlagBits
QUERY_PIPELINE_STATISTIC_TESSELLATION_EVALUATION_SHADER_INVOCATIONS_BIT specifies that queries managed by the pool will count the number of invocations of the tessellation evaluation shader. The counter’s value is incremented each time the tessellation evaluation shader is invoked.

pattern QUERY_PIPELINE_STATISTIC_COMPUTE_SHADER_INVOCATIONS_BIT :: QueryPipelineStatisticFlagBits
QUERY_PIPELINE_STATISTIC_COMPUTE_SHADER_INVOCATIONS_BIT specifies that queries managed by the pool will count the number of compute shader invocations. The counter’s value is incremented every time the compute shader is invoked. Implementations may skip the execution of certain compute shader invocations or execute additional compute shader invocations for implementation-dependent reasons as long as the results of rendering otherwise remain unchanged.

type QueryPipelineStatisticFlags = QueryPipelineStatisticFlagBits