Vulkan.Core10.QueueSemaphore

Documentation

createSemaphore

Arguments

:: forall a io. (Extendss SemaphoreCreateInfo a, PokeChain a, MonadIO io)

=> Device
device is the logical device that creates the semaphore.

-> SemaphoreCreateInfo a
pCreateInfo is a pointer to a SemaphoreCreateInfo structure containing information about how the semaphore is to be created.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io Semaphore

vkCreateSemaphore - Create a new queue semaphore object

Valid Usage (Implicit)

device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid SemaphoreCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pSemaphore must be a valid pointer to a Semaphore handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, Device, Semaphore, SemaphoreCreateInfo

withSemaphore :: forall a io r. (Extendss SemaphoreCreateInfo a, PokeChain a, MonadIO io) => Device -> SemaphoreCreateInfo a -> Maybe AllocationCallbacks -> (io Semaphore -> (Semaphore -> io ()) -> r) -> r

A convenience wrapper to make a compatible pair of calls to createSemaphore and destroySemaphore

To ensure that destroySemaphore is always called: pass bracket (or the allocate function from your favourite resource management library) as the last argument. To just extract the pair pass (,) as the last argument.

destroySemaphore

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that destroys the semaphore.

-> Semaphore
semaphore is the handle of the semaphore to destroy.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io ()

vkDestroySemaphore - Destroy a semaphore object

Valid Usage

All submitted batches that refer to semaphore must have completed execution

If AllocationCallbacks were provided when semaphore was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when semaphore was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid Device handle

If semaphore is not NULL_HANDLE, semaphore must be a valid Semaphore handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If semaphore is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to semaphore must be externally synchronized

See Also

AllocationCallbacks, Device, Semaphore

data SemaphoreCreateInfo (es :: [Type])

VkSemaphoreCreateInfo - Structure specifying parameters of a newly created semaphore

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_SEMAPHORE_CREATE_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of ExportSemaphoreCreateInfo, ExportSemaphoreWin32HandleInfoKHR, or SemaphoreTypeCreateInfo

The sType value of each struct in the pNext chain must be unique

flags must be 0

See Also

SemaphoreCreateFlags, StructureType, createSemaphore

Constructors

SemaphoreCreateInfo

Fields

next :: Chain es

pNext is NULL or a pointer to a structure extending this structure.

flags :: SemaphoreCreateFlags

flags is reserved for future use.

newtype Semaphore

VkSemaphore - Opaque handle to a semaphore object

See Also

AcquireNextImageInfoKHR, BindSparseInfo, ImportSemaphoreFdInfoKHR, ImportSemaphoreWin32HandleInfoKHR, PresentInfoKHR, SemaphoreGetFdInfoKHR, SemaphoreGetWin32HandleInfoKHR, SemaphoreSignalInfo, SemaphoreWaitInfo, SubmitInfo, acquireNextImageKHR, createSemaphore, destroySemaphore, getSemaphoreCounterValue, getSemaphoreCounterValueKHR

Constructors

Semaphore Word64

newtype SemaphoreCreateFlags

VkSemaphoreCreateFlags - Reserved for future use

Description

SemaphoreCreateFlags is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

SemaphoreCreateInfo

Constructors

SemaphoreCreateFlags Flags