Vulkan.Core11.DeviceInitialization

Documentation

enumerateInstanceVersion :: forall io. MonadIO io => io ("apiVersion" ::: Word32)

vkEnumerateInstanceVersion - Query instance-level version before instance creation

Description

Note

The intended behaviour of enumerateInstanceVersion is that an implementation should not need to perform memory allocations and should unconditionally return SUCCESS. The loader, and any enabled layers, may return ERROR_OUT_OF_HOST_MEMORY in the case of a failed memory allocation.

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

See Also

No cross-references are available