Vulkan.Core11.Enums.ExternalFenceFeatureFlagBits

Documentation

newtype ExternalFenceFeatureFlagBits

VkExternalFenceFeatureFlagBits - Bitfield describing features of an external fence handle type

See Also

ExternalFenceFeatureFlags

Constructors

ExternalFenceFeatureFlagBits Flags

Bundled Patterns

pattern EXTERNAL_FENCE_FEATURE_EXPORTABLE_BIT :: ExternalFenceFeatureFlagBits
EXTERNAL_FENCE_FEATURE_EXPORTABLE_BIT specifies handles of this type can be exported from Vulkan fence objects.

pattern EXTERNAL_FENCE_FEATURE_IMPORTABLE_BIT :: ExternalFenceFeatureFlagBits
EXTERNAL_FENCE_FEATURE_IMPORTABLE_BIT specifies handles of this type can be imported to Vulkan fence objects.

type ExternalFenceFeatureFlags = ExternalFenceFeatureFlagBits