Vulkan.Core11.Enums.ExternalMemoryFeatureFlagBits

Documentation

type ExternalMemoryFeatureFlags = ExternalMemoryFeatureFlagBits

newtype ExternalMemoryFeatureFlagBits

VkExternalMemoryFeatureFlagBits - Bitmask specifying features of an external memory handle type

Description

Because their semantics in external APIs roughly align with that of an image or buffer with a dedicated allocation in Vulkan, implementations are required to report EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT for the following external handle types:

Implementations must not report EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT for buffers with external handle type EXTERNAL_MEMORY_HANDLE_TYPE_ANDROID_HARDWARE_BUFFER_BIT_ANDROID. Implementations must not report EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT for images or buffers with external handle type EXTERNAL_MEMORY_HANDLE_TYPE_HOST_ALLOCATION_BIT_EXT, or EXTERNAL_MEMORY_HANDLE_TYPE_HOST_MAPPED_FOREIGN_MEMORY_BIT_EXT.

See Also

ExternalMemoryFeatureFlags

Constructors

ExternalMemoryFeatureFlagBits Flags

Bundled Patterns

pattern EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT :: ExternalMemoryFeatureFlagBits
EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT specifies that images or buffers created with the specified parameters and handle type must use the mechanisms defined by MemoryDedicatedRequirements and MemoryDedicatedAllocateInfo to create (or import) a dedicated allocation for the image or buffer.

pattern EXTERNAL_MEMORY_FEATURE_EXPORTABLE_BIT :: ExternalMemoryFeatureFlagBits
EXTERNAL_MEMORY_FEATURE_EXPORTABLE_BIT specifies that handles of this type can be exported from Vulkan memory objects.

pattern EXTERNAL_MEMORY_FEATURE_IMPORTABLE_BIT :: ExternalMemoryFeatureFlagBits
EXTERNAL_MEMORY_FEATURE_IMPORTABLE_BIT specifies that handles of this type can be imported as Vulkan memory objects.