Vulkan.Core11.Enums.SamplerYcbcrRange

Documentation

newtype SamplerYcbcrRange

VkSamplerYcbcrRange - Range of encoded values in a color space

Description

The formulae for these conversions is described in the Sampler Y′CBCR Range Expansion section of the Image Operations chapter.

No range modification takes place if ycbcrModel is SAMPLER_YCBCR_MODEL_CONVERSION_RGB_IDENTITY; the ycbcrRange field of SamplerYcbcrConversionCreateInfo is ignored in this case.

See Also

AndroidHardwareBufferFormatPropertiesANDROID, SamplerYcbcrConversionCreateInfo

Constructors

SamplerYcbcrRange Int32

Bundled Patterns

pattern SAMPLER_YCBCR_RANGE_ITU_FULL :: SamplerYcbcrRange
SAMPLER_YCBCR_RANGE_ITU_FULL specifies that the full range of the encoded values are valid and interpreted according to the ITU “full range” quantization rules.

pattern SAMPLER_YCBCR_RANGE_ITU_NARROW :: SamplerYcbcrRange
SAMPLER_YCBCR_RANGE_ITU_NARROW specifies that headroom and foot room are reserved in the numerical range of encoded values, and the remaining values are expanded according to the ITU “narrow range” quantization rules.