Vulkan.Core11.Enums.TessellationDomainOrigin

Documentation

newtype TessellationDomainOrigin

VkTessellationDomainOrigin - Enum describing tessellation domain origin

Description

This enum affects how the VertexOrderCw and VertexOrderCcw tessellation execution modes are interpreted, since the winding is defined relative to the orientation of the domain.

See Also

PipelineTessellationDomainOriginStateCreateInfo

Constructors

TessellationDomainOrigin Int32

Bundled Patterns

pattern TESSELLATION_DOMAIN_ORIGIN_UPPER_LEFT :: TessellationDomainOrigin
TESSELLATION_DOMAIN_ORIGIN_UPPER_LEFT specifies that the origin of the domain space is in the upper left corner, as shown in figure https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#img-tessellation-topology-ul.

pattern TESSELLATION_DOMAIN_ORIGIN_LOWER_LEFT :: TessellationDomainOrigin
TESSELLATION_DOMAIN_ORIGIN_LOWER_LEFT specifies that the origin of the domain space is in the lower left corner, as shown in figure https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#img-tessellation-topology-ll.