Vulkan.Core12.Enums.DriverId

Documentation

newtype DriverId

VkDriverId - Khronos driver IDs

Description

Note

Khronos driver IDs may be allocated by vendors at any time. There may be multiple driver IDs for the same vendor, representing different drivers (for e.g. different platforms, proprietary or open source, etc.). Only the latest canonical versions of this Specification, of the corresponding vk.xml API Registry, and of the corresponding vulkan_core.h header file must contain all reserved Khronos driver IDs.

Only driver IDs registered with Khronos are given symbolic names. There may be unregistered driver IDs returned.

See Also

PhysicalDeviceDriverProperties, PhysicalDeviceVulkan12Properties

Constructors

DriverId Int32

Bundled Patterns

pattern DRIVER_ID_AMD_PROPRIETARY :: DriverId

pattern DRIVER_ID_AMD_OPEN_SOURCE :: DriverId

pattern DRIVER_ID_MESA_RADV :: DriverId

pattern DRIVER_ID_NVIDIA_PROPRIETARY :: DriverId

pattern DRIVER_ID_INTEL_PROPRIETARY_WINDOWS :: DriverId

pattern DRIVER_ID_INTEL_OPEN_SOURCE_MESA :: DriverId

pattern DRIVER_ID_IMAGINATION_PROPRIETARY :: DriverId

pattern DRIVER_ID_QUALCOMM_PROPRIETARY :: DriverId

pattern DRIVER_ID_ARM_PROPRIETARY :: DriverId

pattern DRIVER_ID_GOOGLE_SWIFTSHADER :: DriverId

pattern DRIVER_ID_GGP_PROPRIETARY :: DriverId

pattern DRIVER_ID_BROADCOM_PROPRIETARY :: DriverId

pattern DRIVER_ID_MESA_LLVMPIPE :: DriverId