Vulkan.Extensions.VK_AMD_draw_indirect_count

Description

Name

VK_AMD_draw_indirect_count - device extension

VK_AMD_draw_indirect_count

Name String

VK_AMD_draw_indirect_count

Extension Type

Device extension

Registered Extension Number

34

Revision

2

Extension and Version Dependencies

Requires Vulkan 1.0

Deprecation state

Promoted to VK_KHR_draw_indirect_count extension

Which in turn was promoted to Vulkan 1.2

Contact

Daniel Rakos https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_AMD_draw_indirect_count:%20&body=@drakos-amd%20

Other Extension Metadata

Last Modified Date

2016-08-23

Interactions and External Dependencies

Promoted to VK_KHR_draw_indirect_count

IP Status

No known IP claims.

Contributors

Matthaeus G. Chajdas, AMD

Derrick Owens, AMD

Graham Sellers, AMD

Daniel Rakos, AMD

Dominik Witczak, AMD

Description

This extension allows an application to source the number of draw calls for indirect draw calls from a buffer. This enables applications to generate arbitrary amounts of draw commands and execute them without host intervention.

Promotion to VK_KHR_draw_indirect_count

All functionality in this extension is included in VK_KHR_draw_indirect_count, with the suffix changed to KHR. The original type, enum and command names are still available as aliases of the core functionality.

New Commands

cmdDrawIndexedIndirectCountAMD

cmdDrawIndirectCountAMD

New Enum Constants

AMD_DRAW_INDIRECT_COUNT_EXTENSION_NAME

AMD_DRAW_INDIRECT_COUNT_SPEC_VERSION

Version History

Revision 2, 2016-08-23 (Dominik Witczak)

Minor fixes

Revision 1, 2016-07-21 (Matthaeus Chajdas)

Initial draft

See Also

cmdDrawIndexedIndirectCountAMD, cmdDrawIndirectCountAMD

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

cmdDrawIndirectCountAMD :: MonadIO io => CommandBuffer -> Buffer -> ("offset" ::: DeviceSize) -> ("countBuffer" ::: Buffer) -> ("countBufferOffset" ::: DeviceSize) -> ("maxDrawCount" ::: Word32) -> ("stride" ::: Word32) -> io ()

cmdDrawIndexedIndirectCountAMD :: MonadIO io => CommandBuffer -> Buffer -> ("offset" ::: DeviceSize) -> ("countBuffer" ::: Buffer) -> ("countBufferOffset" ::: DeviceSize) -> ("maxDrawCount" ::: Word32) -> ("stride" ::: Word32) -> io ()

type AMD_DRAW_INDIRECT_COUNT_SPEC_VERSION = 2

pattern AMD_DRAW_INDIRECT_COUNT_SPEC_VERSION :: forall a. Integral a => a

type AMD_DRAW_INDIRECT_COUNT_EXTENSION_NAME = "VK_AMD_draw_indirect_count"

pattern AMD_DRAW_INDIRECT_COUNT_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a