Vulkan.Extensions.VK_AMD_mixed_attachment_samples

Description

Name

VK_AMD_mixed_attachment_samples - device extension

VK_AMD_mixed_attachment_samples

Name String

VK_AMD_mixed_attachment_samples

Extension Type

Device extension

Registered Extension Number

137

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Contact

Matthaeus G. Chajdas https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_AMD_mixed_attachment_samples:%20&body=@anteru%20

Other Extension Metadata

Last Modified Date

2017-07-24

Contributors

Mais Alnasser, AMD

Matthaeus G. Chajdas, AMD

Maciej Jesionowski, AMD

Daniel Rakos, AMD

Description

This extension enables applications to use multisampled rendering with a depth/stencil sample count that is larger than the color sample count. Having a depth/stencil sample count larger than the color sample count allows maintaining geometry and coverage information at a higher sample rate than color information. All samples are depth/stencil tested, but only the first color sample count number of samples get a corresponding color output.

New Enum Constants

AMD_MIXED_ATTACHMENT_SAMPLES_EXTENSION_NAME

AMD_MIXED_ATTACHMENT_SAMPLES_SPEC_VERSION

Issues

None.

Version History

Revision 1, 2017-07-24 (Daniel Rakos)

Internal revisions

See Also

No cross-references are available

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

type AMD_MIXED_ATTACHMENT_SAMPLES_SPEC_VERSION = 1

pattern AMD_MIXED_ATTACHMENT_SAMPLES_SPEC_VERSION :: forall a. Integral a => a

type AMD_MIXED_ATTACHMENT_SAMPLES_EXTENSION_NAME = "VK_AMD_mixed_attachment_samples"

pattern AMD_MIXED_ATTACHMENT_SAMPLES_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a