Vulkan.Extensions.VK_AMD_shader_core_properties2

Documentation

data PhysicalDeviceShaderCoreProperties2AMD

VkPhysicalDeviceShaderCoreProperties2AMD - Structure describing shader core properties that can be supported by an implementation

Members

The members of the PhysicalDeviceShaderCoreProperties2AMD structure describe the following implementation-dependent limits:

Description

If the PhysicalDeviceShaderCoreProperties2AMD structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

ShaderCorePropertiesFlagsAMD, StructureType

Constructors

PhysicalDeviceShaderCoreProperties2AMD

Fields

shaderCoreFeatures :: ShaderCorePropertiesFlagsAMD

shaderCoreFeatures is a bitmask of ShaderCorePropertiesFlagBitsAMD indicating the set of features supported by the shader core.

activeComputeUnitCount :: Word32

activeComputeUnitCount is an unsigned integer value indicating the number of compute units that have been enabled.

newtype ShaderCorePropertiesFlagBitsAMD

VkShaderCorePropertiesFlagBitsAMD - Bitmask specifying shader core properties

See Also

PhysicalDeviceShaderCoreProperties2AMD, ShaderCorePropertiesFlagsAMD

Constructors

ShaderCorePropertiesFlagBitsAMD Flags

type ShaderCorePropertiesFlagsAMD = ShaderCorePropertiesFlagBitsAMD

type AMD_SHADER_CORE_PROPERTIES_2_SPEC_VERSION = 1

pattern AMD_SHADER_CORE_PROPERTIES_2_SPEC_VERSION :: forall a. Integral a => a

type AMD_SHADER_CORE_PROPERTIES_2_EXTENSION_NAME = "VK_AMD_shader_core_properties2"

pattern AMD_SHADER_CORE_PROPERTIES_2_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a