Vulkan.Extensions.VK_AMD_texture_gather_bias_lod

Description

Name

VK_AMD_texture_gather_bias_lod - device extension

VK_AMD_texture_gather_bias_lod

Name String

VK_AMD_texture_gather_bias_lod

Extension Type

Device extension

Registered Extension Number

42

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_get_physical_device_properties2

Contact

Rex Xu https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_AMD_texture_gather_bias_lod:%20&body=@amdrexu%20

Other Extension Metadata

Last Modified Date

2017-03-21

IP Status

No known IP claims.

Interactions and External Dependencies

This extension requires SPV_AMD_texture_gather_bias_lod

Contributors

Dominik Witczak, AMD

Daniel Rakos, AMD

Graham Sellers, AMD

Matthaeus G. Chajdas, AMD

Qun Lin, AMD

Rex Xu, AMD

Timothy Lottes, AMD

Description

This extension adds two related features.

Firstly, support for the following SPIR-V extension in Vulkan is added:

SPV_AMD_texture_gather_bias_lod

Secondly, the extension allows the application to query which formats can be used together with the new function prototypes introduced by the SPIR-V extension.

New Structures

Extending ImageFormatProperties2:

TextureLODGatherFormatPropertiesAMD

New Enum Constants

AMD_TEXTURE_GATHER_BIAS_LOD_EXTENSION_NAME

AMD_TEXTURE_GATHER_BIAS_LOD_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_TEXTURE_LOD_GATHER_FORMAT_PROPERTIES_AMD

New SPIR-V Capabilities

ImageGatherBiasLodAMD

Examples

struct VkTextureLODGatherFormatPropertiesAMD
{
    VkStructureType sType;
    const void*     pNext;
    VkBool32        supportsTextureGatherLODBiasAMD;
};

// ----------------------------------------------------------------------------------------
// How to detect if an image format can be used with the new function prototypes.
VkPhysicalDeviceImageFormatInfo2   formatInfo;
VkImageFormatProperties2           formatProps;
VkTextureLODGatherFormatPropertiesAMD textureLODGatherSupport;

textureLODGatherSupport.sType = VK_STRUCTURE_TYPE_TEXTURE_LOD_GATHER_FORMAT_PROPERTIES_AMD;
textureLODGatherSupport.pNext = nullptr;

formatInfo.sType  = VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGE_FORMAT_INFO_2;
formatInfo.pNext  = nullptr;
formatInfo.format = ...;
formatInfo.type   = ...;
formatInfo.tiling = ...;
formatInfo.usage  = ...;
formatInfo.flags  = ...;

formatProps.sType = VK_STRUCTURE_TYPE_IMAGE_FORMAT_PROPERTIES_2;
formatProps.pNext = &textureLODGatherSupport;

vkGetPhysicalDeviceImageFormatProperties2(physical_device, &formatInfo, &formatProps);

if (textureLODGatherSupport.supportsTextureGatherLODBiasAMD == VK_TRUE)
{
    // physical device supports SPV_AMD_texture_gather_bias_lod for the specified
    // format configuration.
}
else
{
    // physical device does not support SPV_AMD_texture_gather_bias_lod for the
    // specified format configuration.
}

Version History

Revision 1, 2017-03-21 (Dominik Witczak)

Initial draft

See Also

TextureLODGatherFormatPropertiesAMD

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data TextureLODGatherFormatPropertiesAMD

VkTextureLODGatherFormatPropertiesAMD - Structure informing whether or not texture gather bias/LOD functionality is supported for a given image format and a given physical device.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

TextureLODGatherFormatPropertiesAMD

Fields

supportsTextureGatherLODBiasAMD :: Bool

supportsTextureGatherLODBiasAMD tells if the image format can be used with texture gather bias/LOD functions, as introduced by the VK_AMD_texture_gather_bias_lod extension. This field is set by the implementation. User-specified value is ignored.

type AMD_TEXTURE_GATHER_BIAS_LOD_SPEC_VERSION = 1

pattern AMD_TEXTURE_GATHER_BIAS_LOD_SPEC_VERSION :: forall a. Integral a => a

type AMD_TEXTURE_GATHER_BIAS_LOD_EXTENSION_NAME = "VK_AMD_texture_gather_bias_lod"

pattern AMD_TEXTURE_GATHER_BIAS_LOD_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a