Vulkan.Extensions.VK_EXT_acquire_xlib_display

Documentation

acquireXlibDisplayEXT

Arguments

:: forall io. MonadIO io

=> PhysicalDevice
physicalDevice The physical device the display is on.

physicalDevice must be a valid PhysicalDevice handle

-> ("dpy" ::: Ptr Display)
dpy A connection to the X11 server that currently owns display.

dpy must be a valid pointer to a Display value

-> DisplayKHR
display The display the caller wishes to control in Vulkan.

display must be a valid DisplayKHR handle

display must have been created, allocated, or retrieved from physicalDevice

-> io ()

vkAcquireXlibDisplayEXT - Acquire access to a VkDisplayKHR using Xlib

Description

All permissions necessary to control the display are granted to the Vulkan instance associated with physicalDevice until the display is released or the X11 connection specified by dpy is terminated. Permission to access the display may be temporarily revoked during periods when the X11 server from which control was acquired itself loses access to display. During such periods, operations which require access to the display must fail with an approriate error code. If the X11 server associated with dpy does not own display, or if permission to access it has already been acquired by another entity, the call must return the error code ERROR_INITIALIZATION_FAILED.

Note

One example of when an X11 server loses access to a display is when it loses ownership of its virtual terminal.

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_INITIALIZATION_FAILED

See Also

DisplayKHR, PhysicalDevice

getRandROutputDisplayEXT

Arguments

:: forall io. MonadIO io

=> PhysicalDevice
physicalDevice The physical device to query the display handle on.

physicalDevice must be a valid PhysicalDevice handle

-> ("dpy" ::: Ptr Display)
dpy A connection to the X11 server from which rrOutput was queried.

dpy must be a valid pointer to a Display value

-> RROutput
rrOutput An X11 RandR output ID.

-> io DisplayKHR

vkGetRandROutputDisplayEXT - Query the VkDisplayKHR corresponding to an X11 RandR Output

Description

If there is no DisplayKHR corresponding to rrOutput on physicalDevice, NULL_HANDLE must be returned in pDisplay.

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

See Also

DisplayKHR, PhysicalDevice

type EXT_ACQUIRE_XLIB_DISPLAY_SPEC_VERSION = 1

pattern EXT_ACQUIRE_XLIB_DISPLAY_SPEC_VERSION :: forall a. Integral a => a

type EXT_ACQUIRE_XLIB_DISPLAY_EXTENSION_NAME = "VK_EXT_acquire_xlib_display"

pattern EXT_ACQUIRE_XLIB_DISPLAY_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

type RROutput = Word64

newtype DisplayKHR

VkDisplayKHR - Opaque handle to a display object

See Also

DisplayPlanePropertiesKHR, DisplayPropertiesKHR, acquireXlibDisplayEXT, createDisplayModeKHR, displayPowerControlEXT, getDisplayModeProperties2KHR, getDisplayModePropertiesKHR, getDisplayPlaneSupportedDisplaysKHR, getRandROutputDisplayEXT, registerDisplayEventEXT, releaseDisplayEXT

Constructors

DisplayKHR Word64

type Display = Ptr ()