Vulkan.Extensions.VK_EXT_astc_decode_mode

Description

Name

VK_EXT_astc_decode_mode - device extension

VK_EXT_astc_decode_mode

Name String

VK_EXT_astc_decode_mode

Extension Type

Device extension

Registered Extension Number

68

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_get_physical_device_properties2

Contact

Jan-Harald Fredriksen https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_EXT_astc_decode_mode:%20&body=@janharaldfredriksen-arm%20

Other Extension Metadata

Last Modified Date

2018-08-07

Contributors

Jan-Harald Fredriksen, Arm

Description

The existing specification requires that low dynamic range (LDR) ASTC textures are decompressed to FP16 values per component. In many cases, decompressing LDR textures to a lower precision intermediate result gives acceptable image quality. Source material for LDR textures is typically authored as 8-bit UNORM values, so decoding to FP16 values adds little value. On the other hand, reducing precision of the decoded result reduces the size of the decompressed data, potentially improving texture cache performance and saving power.

The goal of this extension is to enable this efficiency gain on existing ASTC texture data. This is achieved by giving the application the ability to select the intermediate decoding precision.

Three decoding options are provided:

Decode to FORMAT_R16G16B16A16_SFLOAT precision: This is the default, and matches the required behavior in the core API.

Decode to FORMAT_R8G8B8A8_UNORM precision: This is provided as an option in LDR mode.

Decode to FORMAT_E5B9G9R9_UFLOAT_PACK32 precision: This is provided as an option in both LDR and HDR mode. In this mode, negative values cannot be represented and are clamped to zero. The alpha component is ignored, and the results are as if alpha was 1.0. This decode mode is optional and support can be queried via the physical device properties.

New Structures

Extending ImageViewCreateInfo:

ImageViewASTCDecodeModeEXT

Extending PhysicalDeviceFeatures2, DeviceCreateInfo:

PhysicalDeviceASTCDecodeFeaturesEXT

New Enum Constants

EXT_ASTC_DECODE_MODE_EXTENSION_NAME

EXT_ASTC_DECODE_MODE_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_IMAGE_VIEW_ASTC_DECODE_MODE_EXT

STRUCTURE_TYPE_PHYSICAL_DEVICE_ASTC_DECODE_FEATURES_EXT

Issues

1) Are implementations allowed to decode at a higher precision than what is requested?

RESOLUTION: No.
If we allow this, then this extension could be exposed on all
implementations that support ASTC.
But developers would have no way of knowing what precision was actually
used, and thus whether the image quality is sufficient at reduced
precision.

2) Should the decode mode be image view state and/or sampler state?

RESOLUTION: Image view state only.
Some implementations treat the different decode modes as different
texture formats.

Example

Create an image view that decodes to FORMAT_R8G8B8A8_UNORM precision:

    VkImageViewASTCDecodeModeEXT decodeMode =
    {
        VK_STRUCTURE_TYPE_IMAGE_VIEW_ASTC_DECODE_MODE_EXT, // sType
        NULL, // pNext
        VK_FORMAT_R8G8B8A8_UNORM // decode mode
    };

    VkImageViewCreateInfo createInfo =
    {
        VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO, // sType
        &decodeMode, // pNext
        // flags, image, viewType set to application-desired values
        VK_FORMAT_ASTC_8x8_UNORM_BLOCK, // format
        // components, subresourceRange set to application-desired values
    };

    VkImageView imageView;
    VkResult result = vkCreateImageView(
        device,
        &createInfo,
        NULL,
        &imageView);

Version History

Revision 1, 2018-08-07 (Jan-Harald Fredriksen)

Initial revision

See Also

ImageViewASTCDecodeModeEXT, PhysicalDeviceASTCDecodeFeaturesEXT

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data ImageViewASTCDecodeModeEXT

VkImageViewASTCDecodeModeEXT - Structure describing the ASTC decode mode for an image view

Valid Usage

decodeMode must be one of FORMAT_R16G16B16A16_SFLOAT, FORMAT_R8G8B8A8_UNORM, or FORMAT_E5B9G9R9_UFLOAT_PACK32

If the decodeModeSharedExponent feature is not enabled, decodeMode must not be FORMAT_E5B9G9R9_UFLOAT_PACK32

If decodeMode is FORMAT_R8G8B8A8_UNORM the image view must not include blocks using any of the ASTC HDR modes

format of the image view must be one of the ASTC Compressed Image Formats

If format uses sRGB encoding then the decodeMode has no effect.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_IMAGE_VIEW_ASTC_DECODE_MODE_EXT

decodeMode must be a valid Format value

See Also

Format, StructureType

Constructors

ImageViewASTCDecodeModeEXT

Fields

decodeMode :: Format

decodeMode is the intermediate format used to decode ASTC compressed formats.

data PhysicalDeviceASTCDecodeFeaturesEXT

VkPhysicalDeviceASTCDecodeFeaturesEXT - Structure describing ASTC decode mode features

Members

The members of the PhysicalDeviceASTCDecodeFeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceASTCDecodeFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether each feature is supported. PhysicalDeviceASTCDecodeFeaturesEXT can also be included in the pNext chain of createDevice to enable features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceASTCDecodeFeaturesEXT

Fields

decodeModeSharedExponent :: Bool

decodeModeSharedExponent indicates whether the implementation supports decoding ASTC compressed formats to FORMAT_E5B9G9R9_UFLOAT_PACK32 internal precision.

type EXT_ASTC_DECODE_MODE_SPEC_VERSION = 1

pattern EXT_ASTC_DECODE_MODE_SPEC_VERSION :: forall a. Integral a => a

type EXT_ASTC_DECODE_MODE_EXTENSION_NAME = "VK_EXT_astc_decode_mode"

pattern EXT_ASTC_DECODE_MODE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a