Vulkan.Extensions.VK_EXT_blend_operation_advanced

Documentation

data PhysicalDeviceBlendOperationAdvancedFeaturesEXT

VkPhysicalDeviceBlendOperationAdvancedFeaturesEXT - Structure describing advanced blending features that can be supported by an implementation

Members

The members of the PhysicalDeviceBlendOperationAdvancedFeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceBlendOperationAdvancedFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether each feature is supported. PhysicalDeviceBlendOperationAdvancedFeaturesEXT can also be included in the pNext chain of DeviceCreateInfo to enable the features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceBlendOperationAdvancedFeaturesEXT

Fields

advancedBlendCoherentOperations :: Bool

advancedBlendCoherentOperations specifies whether blending using advanced blend operations is guaranteed to execute atomically and in primitive order. If this is TRUE, ACCESS_COLOR_ATTACHMENT_READ_NONCOHERENT_BIT_EXT is treated the same as ACCESS_COLOR_ATTACHMENT_READ_BIT, and advanced blending needs no additional synchronization over basic blending. If this is FALSE, then memory dependencies are required to guarantee order between two advanced blending operations that occur on the same sample.

data PhysicalDeviceBlendOperationAdvancedPropertiesEXT

VkPhysicalDeviceBlendOperationAdvancedPropertiesEXT - Structure describing advanced blending limits that can be supported by an implementation

Members

The members of the PhysicalDeviceBlendOperationAdvancedPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the PhysicalDeviceBlendOperationAdvancedPropertiesEXT structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceBlendOperationAdvancedPropertiesEXT

Fields

advancedBlendMaxColorAttachments :: Word32

advancedBlendMaxColorAttachments is one greater than the highest color attachment index that can be used in a subpass, for a pipeline that uses an advanced blend operation.

advancedBlendIndependentBlend :: Bool

advancedBlendIndependentBlend specifies whether advanced blend operations can vary per-attachment.

advancedBlendNonPremultipliedSrcColor :: Bool

advancedBlendNonPremultipliedSrcColor specifies whether the source color can be treated as non-premultiplied. If this is FALSE, then PipelineColorBlendAdvancedStateCreateInfoEXT::srcPremultiplied must be TRUE.

advancedBlendNonPremultipliedDstColor :: Bool

advancedBlendNonPremultipliedDstColor specifies whether the destination color can be treated as non-premultiplied. If this is FALSE, then PipelineColorBlendAdvancedStateCreateInfoEXT::dstPremultiplied must be TRUE.

advancedBlendCorrelatedOverlap :: Bool

advancedBlendCorrelatedOverlap specifies whether the overlap mode can be treated as correlated. If this is FALSE, then PipelineColorBlendAdvancedStateCreateInfoEXT::blendOverlap must be BLEND_OVERLAP_UNCORRELATED_EXT.

advancedBlendAllOperations :: Bool

advancedBlendAllOperations specifies whether all advanced blend operation enums are supported. See the valid usage of PipelineColorBlendAttachmentState.

data PipelineColorBlendAdvancedStateCreateInfoEXT

VkPipelineColorBlendAdvancedStateCreateInfoEXT - Structure specifying parameters that affect advanced blend operations

Description

If this structure is not present, srcPremultiplied and dstPremultiplied are both considered to be TRUE, and blendOverlap is considered to be BLEND_OVERLAP_UNCORRELATED_EXT.

Valid Usage

If the non-premultiplied source color property is not supported, srcPremultiplied must be TRUE

If the non-premultiplied destination color property is not supported, dstPremultiplied must be TRUE

If the correlated overlap property is not supported, blendOverlap must be BLEND_OVERLAP_UNCORRELATED_EXT

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_ADVANCED_STATE_CREATE_INFO_EXT

blendOverlap must be a valid BlendOverlapEXT value

See Also

BlendOverlapEXT, Bool32, StructureType

Constructors

PipelineColorBlendAdvancedStateCreateInfoEXT

Fields

srcPremultiplied :: Bool

srcPremultiplied specifies whether the source color of the blend operation is treated as premultiplied.

dstPremultiplied :: Bool

dstPremultiplied specifies whether the destination color of the blend operation is treated as premultiplied.

blendOverlap :: BlendOverlapEXT

blendOverlap is a BlendOverlapEXT value specifying how the source and destination sample’s coverage is correlated.

newtype BlendOverlapEXT

VkBlendOverlapEXT - Enumerant specifying the blend overlap parameter

Description

'

Overlap Mode 	 Weighting Equations

BLEND_OVERLAP_UNCORRELATED_EXT 	 \[ \begin{aligned} p_0(A_s,A_d) & = A_sA_d \\ p_1(A_s,A_d) & = A_s(1-A_d) \\ p_2(A_s,A_d) & = A_d(1-A_s) \\ \end{aligned}\]

BLEND_OVERLAP_CONJOINT_EXT 	 \[ \begin{aligned} p_0(A_s,A_d) & = min(A_s,A_d) \\ p_1(A_s,A_d) & = max(A_s-A_d,0) \\ p_2(A_s,A_d) & = max(A_d-A_s,0) \\ \end{aligned}\]

BLEND_OVERLAP_DISJOINT_EXT 	 \[ \begin{aligned} p_0(A_s,A_d) & = max(A_s+A_d-1,0) \\ p_1(A_s,A_d) & = min(A_s,1-A_d) \\ p_2(A_s,A_d) & = min(A_d,1-A_s) \\ \end{aligned}\]

Advanced Blend Overlap Modes

See Also

PipelineColorBlendAdvancedStateCreateInfoEXT

Constructors

BlendOverlapEXT Int32

Bundled Patterns

pattern BLEND_OVERLAP_UNCORRELATED_EXT :: BlendOverlapEXT
BLEND_OVERLAP_UNCORRELATED_EXT specifies that there is no correlation between the source and destination coverage.

pattern BLEND_OVERLAP_DISJOINT_EXT :: BlendOverlapEXT
BLEND_OVERLAP_DISJOINT_EXT specifies that the source and destination coverage are considered to have minimal overlap.

pattern BLEND_OVERLAP_CONJOINT_EXT :: BlendOverlapEXT
BLEND_OVERLAP_CONJOINT_EXT specifies that the source and destination coverage are considered to have maximal overlap.

type EXT_BLEND_OPERATION_ADVANCED_SPEC_VERSION = 2

pattern EXT_BLEND_OPERATION_ADVANCED_SPEC_VERSION :: forall a. Integral a => a

type EXT_BLEND_OPERATION_ADVANCED_EXTENSION_NAME = "VK_EXT_blend_operation_advanced"

pattern EXT_BLEND_OPERATION_ADVANCED_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a