Vulkan.Extensions.VK_EXT_conditional_rendering

Documentation

cmdBeginConditionalRenderingEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which this command will be recorded.

-> ConditionalRenderingBeginInfoEXT
pConditionalRenderingBegin is a pointer to a ConditionalRenderingBeginInfoEXT structure specifying parameters of conditional rendering.

-> io ()

vkCmdBeginConditionalRenderingEXT - Define the beginning of a conditional rendering block

Valid Usage

Conditional rendering must not already be active

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pConditionalRenderingBegin must be a valid pointer to a valid ConditionalRenderingBeginInfoEXT structure

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute

See Also

CommandBuffer, ConditionalRenderingBeginInfoEXT

cmdUseConditionalRenderingEXT :: forall io r. MonadIO io => CommandBuffer -> ConditionalRenderingBeginInfoEXT -> io r -> io r

This function will call the supplied action between calls to cmdBeginConditionalRenderingEXT and cmdEndConditionalRenderingEXT

Note that cmdEndConditionalRenderingEXT is *not* called if an exception is thrown by the inner action.

cmdEndConditionalRenderingEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which this command will be recorded.

-> io ()

vkCmdEndConditionalRenderingEXT - Define the end of a conditional rendering block

Description

Once ended, conditional rendering becomes inactive.

Valid Usage

Conditional rendering must be active

If conditional rendering was made active outside of a render pass instance, it must not be ended inside a render pass instance

If conditional rendering was made active within a subpass it must be ended in the same subpass

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute

See Also

CommandBuffer

data ConditionalRenderingBeginInfoEXT

VkConditionalRenderingBeginInfoEXT - Structure specifying conditional rendering begin info

Description

If the 32-bit value at offset in buffer memory is zero, then the rendering commands are discarded, otherwise they are executed as normal. If the value of the predicate in buffer memory changes while conditional rendering is active, the rendering commands may be discarded in an implementation-dependent way. Some implementations may latch the value of the predicate upon beginning conditional rendering while others may read it before every rendering command.

Valid Usage

If buffer is non-sparse then it must be bound completely and contiguously to a single DeviceMemory object

buffer must have been created with the BUFFER_USAGE_CONDITIONAL_RENDERING_BIT_EXT bit set

offset must be less than the size of buffer by at least 32 bits

offset must be a multiple of 4

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_CONDITIONAL_RENDERING_BEGIN_INFO_EXT

pNext must be NULL

buffer must be a valid Buffer handle

flags must be a valid combination of ConditionalRenderingFlagBitsEXT values

See Also

Buffer, ConditionalRenderingFlagsEXT, DeviceSize, StructureType, cmdBeginConditionalRenderingEXT

Constructors

ConditionalRenderingBeginInfoEXT

Fields

buffer :: Buffer

buffer is a buffer containing the predicate for conditional rendering.

offset :: DeviceSize

offset is the byte offset into buffer where the predicate is located.

flags :: ConditionalRenderingFlagsEXT

flags is a bitmask of ConditionalRenderingFlagsEXT specifying the behavior of conditional rendering.

data CommandBufferInheritanceConditionalRenderingInfoEXT

VkCommandBufferInheritanceConditionalRenderingInfoEXT - Structure specifying command buffer inheritance info

Description

If this structure is not present, the behavior is as if conditionalRenderingEnable is FALSE.

Valid Usage

If the inherited conditional rendering feature is not enabled, conditionalRenderingEnable must be FALSE

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_COMMAND_BUFFER_INHERITANCE_CONDITIONAL_RENDERING_INFO_EXT

See Also

Bool32, StructureType

Constructors

CommandBufferInheritanceConditionalRenderingInfoEXT

Fields

conditionalRenderingEnable :: Bool

conditionalRenderingEnable specifies whether the command buffer can be executed while conditional rendering is active in the primary command buffer. If this is TRUE, then this command buffer can be executed whether the primary command buffer has active conditional rendering or not. If this is FALSE, then the primary command buffer must not have conditional rendering active.

data PhysicalDeviceConditionalRenderingFeaturesEXT

VkPhysicalDeviceConditionalRenderingFeaturesEXT - Structure describing if a secondary command buffer can be executed if conditional rendering is active in the primary command buffer

Description

If the PhysicalDeviceConditionalRenderingFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating the implementation-dependent behavior. PhysicalDeviceConditionalRenderingFeaturesEXT can also be included in pNext chain of DeviceCreateInfo to enable the features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceConditionalRenderingFeaturesEXT

Fields

conditionalRendering :: Bool

conditionalRendering specifies whether conditional rendering is supported.

inheritedConditionalRendering :: Bool

inheritedConditionalRendering specifies whether a secondary command buffer can be executed while conditional rendering is active in the primary command buffer.

newtype ConditionalRenderingFlagBitsEXT

VkConditionalRenderingFlagBitsEXT - Specify the behavior of conditional rendering

See Also

ConditionalRenderingFlagsEXT

Constructors

ConditionalRenderingFlagBitsEXT Flags

Bundled Patterns

pattern CONDITIONAL_RENDERING_INVERTED_BIT_EXT :: ConditionalRenderingFlagBitsEXT
CONDITIONAL_RENDERING_INVERTED_BIT_EXT specifies the condition used to determine whether to discard rendering commands or not. That is, if the 32-bit predicate read from buffer memory at offset is zero, the rendering commands are not discarded, and if non zero, then they are discarded.

type ConditionalRenderingFlagsEXT = ConditionalRenderingFlagBitsEXT

type EXT_CONDITIONAL_RENDERING_SPEC_VERSION = 2

pattern EXT_CONDITIONAL_RENDERING_SPEC_VERSION :: forall a. Integral a => a

type EXT_CONDITIONAL_RENDERING_EXTENSION_NAME = "VK_EXT_conditional_rendering"

pattern EXT_CONDITIONAL_RENDERING_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a