Vulkan.Extensions.VK_EXT_conservative_rasterization

Documentation

data PhysicalDeviceConservativeRasterizationPropertiesEXT

VkPhysicalDeviceConservativeRasterizationPropertiesEXT - Structure describing conservative raster properties that can be supported by an implementation

Members

The members of the PhysicalDeviceConservativeRasterizationPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the PhysicalDeviceConservativeRasterizationPropertiesEXT structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits and properties.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceConservativeRasterizationPropertiesEXT

Fields

primitiveOverestimationSize :: Float

primitiveOverestimationSize is the size in pixels the generating primitive is increased at each of its edges during conservative rasterization overestimation mode. Even with a size of 0.0, conservative rasterization overestimation rules still apply and if any part of the pixel rectangle is covered by the generating primitive, fragments are generated for the entire pixel. However implementations may make the pixel coverage area even more conservative by increasing the size of the generating primitive.

maxExtraPrimitiveOverestimationSize :: Float

maxExtraPrimitiveOverestimationSize is the maximum size in pixels of extra overestimation the implementation supports in the pipeline state. A value of 0.0 means the implementation does not support any additional overestimation of the generating primitive during conservative rasterization. A value above 0.0 allows the application to further increase the size of the generating primitive during conservative rasterization overestimation.

extraPrimitiveOverestimationSizeGranularity :: Float

extraPrimitiveOverestimationSizeGranularity is the granularity of extra overestimation that can be specified in the pipeline state between 0.0 and maxExtraPrimitiveOverestimationSize inclusive. A value of 0.0 means the implementation can use the smallest representable non-zero value in the screen space pixel fixed-point grid.

primitiveUnderestimation :: Bool

primitiveUnderestimation is true if the implementation supports the CONSERVATIVE_RASTERIZATION_MODE_UNDERESTIMATE_EXT conservative rasterization mode in addition to CONSERVATIVE_RASTERIZATION_MODE_OVERESTIMATE_EXT. Otherwise the implementation only supports CONSERVATIVE_RASTERIZATION_MODE_OVERESTIMATE_EXT.

conservativePointAndLineRasterization :: Bool

conservativePointAndLineRasterization is true if the implementation supports conservative rasterization of point and line primitives as well as triangle primitives. Otherwise the implementation only supports triangle primitives.

degenerateTrianglesRasterized :: Bool

degenerateTrianglesRasterized is false if the implementation culls primitives generated from triangles that become zero area after they are quantized to the fixed-point rasterization pixel grid. degenerateTrianglesRasterized is true if these primitives are not culled and the provoking vertex attributes and depth value are used for the fragments. The primitive area calculation is done on the primitive generated from the clipped triangle if applicable. Zero area primitives are backfacing and the application can enable backface culling if desired.

degenerateLinesRasterized :: Bool

degenerateLinesRasterized is false if the implementation culls lines that become zero length after they are quantized to the fixed-point rasterization pixel grid. degenerateLinesRasterized is true if zero length lines are not culled and the provoking vertex attributes and depth value are used for the fragments.

fullyCoveredFragmentShaderInputVariable :: Bool

fullyCoveredFragmentShaderInputVariable is true if the implementation supports the SPIR-V builtin fragment shader input variable FullyCoveredEXT which specifies that conservative rasterization is enabled and the fragment area is fully covered by the generating primitive.

conservativeRasterizationPostDepthCoverage :: Bool

conservativeRasterizationPostDepthCoverage is true if the implementation supports conservative rasterization with the PostDepthCoverage execution mode enabled. When supported the SampleMask built-in input variable will reflect the coverage after the early per-fragment depth and stencil tests are applied even when conservative rasterization is enabled. Otherwise PostDepthCoverage execution mode must not be used when conservative rasterization is enabled.

data PipelineRasterizationConservativeStateCreateInfoEXT

VkPipelineRasterizationConservativeStateCreateInfoEXT - Structure specifying conservative raster state

Valid Usage (Implicit)

See Also

ConservativeRasterizationModeEXT, PipelineRasterizationConservativeStateCreateFlagsEXT, StructureType

Constructors

PipelineRasterizationConservativeStateCreateInfoEXT

Fields

flags :: PipelineRasterizationConservativeStateCreateFlagsEXT

flags is reserved for future use.

flags must be 0

conservativeRasterizationMode :: ConservativeRasterizationModeEXT

conservativeRasterizationMode is the conservative rasterization mode to use.

conservativeRasterizationMode must be a valid ConservativeRasterizationModeEXT value

extraPrimitiveOverestimationSize :: Float

extraPrimitiveOverestimationSize is the extra size in pixels to increase the generating primitive during conservative rasterization at each of its edges in X and Y equally in screen space beyond the base overestimation specified in PhysicalDeviceConservativeRasterizationPropertiesEXT::primitiveOverestimationSize.

extraPrimitiveOverestimationSize must be in the range of 0.0 to PhysicalDeviceConservativeRasterizationPropertiesEXT::maxExtraPrimitiveOverestimationSize inclusive

newtype PipelineRasterizationConservativeStateCreateFlagsEXT

VkPipelineRasterizationConservativeStateCreateFlagsEXT - Reserved for future use

Description

PipelineRasterizationConservativeStateCreateFlagsEXT is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

PipelineRasterizationConservativeStateCreateInfoEXT

Constructors

PipelineRasterizationConservativeStateCreateFlagsEXT Flags

newtype ConservativeRasterizationModeEXT

VkConservativeRasterizationModeEXT - Specify the conservative rasterization mode

See Also

PipelineRasterizationConservativeStateCreateInfoEXT

Constructors

ConservativeRasterizationModeEXT Int32

Bundled Patterns

pattern CONSERVATIVE_RASTERIZATION_MODE_DISABLED_EXT :: ConservativeRasterizationModeEXT
CONSERVATIVE_RASTERIZATION_MODE_DISABLED_EXT specifies that conservative rasterization is disabled and rasterization proceeds as normal.

pattern CONSERVATIVE_RASTERIZATION_MODE_OVERESTIMATE_EXT :: ConservativeRasterizationModeEXT
CONSERVATIVE_RASTERIZATION_MODE_OVERESTIMATE_EXT specifies that conservative rasterization is enabled in overestimation mode.

pattern CONSERVATIVE_RASTERIZATION_MODE_UNDERESTIMATE_EXT :: ConservativeRasterizationModeEXT
CONSERVATIVE_RASTERIZATION_MODE_UNDERESTIMATE_EXT specifies that conservative rasterization is enabled in underestimation mode.

type EXT_CONSERVATIVE_RASTERIZATION_SPEC_VERSION = 1

pattern EXT_CONSERVATIVE_RASTERIZATION_SPEC_VERSION :: forall a. Integral a => a

type EXT_CONSERVATIVE_RASTERIZATION_EXTENSION_NAME = "VK_EXT_conservative_rasterization"

pattern EXT_CONSERVATIVE_RASTERIZATION_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a