Vulkan.Extensions.VK_EXT_custom_border_color

Description

Name

VK_EXT_custom_border_color - device extension

VK_EXT_custom_border_color

Name String

VK_EXT_custom_border_color

Extension Type

Device extension

Registered Extension Number

288

Revision

12

Extension and Version Dependencies

Requires Vulkan 1.0

Contact

Liam Middlebrook https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_EXT_custom_border_color:%20&body=@liam-middlebrook%20

Other Extension Metadata

Last Modified Date

2020-04-16

IP Status

No known IP claims.

Contributors

Joshua Ashton, Valve

Hans-Kristian Arntzen, Valve

Philip Rebohle, Valve

Liam Middlebrook, NVIDIA

Jeff Bolz, NVIDIA

Tobias Hector, AMD

Jason Ekstrand, Intel

Spencer Fricke, Samsung Electronics

Graeme Leese, Broadcom

Jesse Hall, Google

Jan-Harald Fredriksen, ARM

Tom Olson, ARM

Stuart Smith, Imagination Technologies

Donald Scorgie, Imagination Technologies

Alex Walters, Imagination Technologies

Peter Quayle, Imagination Technologies

Description

This extension provides cross-vendor functionality to specify a custom border color for use when the sampler address mode SAMPLER_ADDRESS_MODE_CLAMP_TO_BORDER is used.

To create a sampler which uses a custom border color set SamplerCreateInfo::borderColor to one of:

BORDER_COLOR_FLOAT_CUSTOM_EXT

BORDER_COLOR_INT_CUSTOM_EXT

When BORDER_COLOR_FLOAT_CUSTOM_EXT or BORDER_COLOR_INT_CUSTOM_EXT is used, applications must provide a SamplerCustomBorderColorCreateInfoEXT in the pNext chain for SamplerCreateInfo.

New Structures

Extending PhysicalDeviceFeatures2, DeviceCreateInfo:

PhysicalDeviceCustomBorderColorFeaturesEXT

Extending PhysicalDeviceProperties2:

PhysicalDeviceCustomBorderColorPropertiesEXT

Extending SamplerCreateInfo:

SamplerCustomBorderColorCreateInfoEXT

New Enum Constants

EXT_CUSTOM_BORDER_COLOR_EXTENSION_NAME

EXT_CUSTOM_BORDER_COLOR_SPEC_VERSION

Extending BorderColor:

BORDER_COLOR_FLOAT_CUSTOM_EXT

BORDER_COLOR_INT_CUSTOM_EXT

Extending StructureType:

STRUCTURE_TYPE_PHYSICAL_DEVICE_CUSTOM_BORDER_COLOR_FEATURES_EXT

STRUCTURE_TYPE_PHYSICAL_DEVICE_CUSTOM_BORDER_COLOR_PROPERTIES_EXT

STRUCTURE_TYPE_SAMPLER_CUSTOM_BORDER_COLOR_CREATE_INFO_EXT

Issues

1) Should VkClearColorValue be used for the border color value, or should we have our own struct/union? Do we need to specify the type of the input values for the components? This is more of a concern if VkClearColorValue is used here because it provides a union of float,int,uint types.

RESOLVED: Will re-use existing VkClearColorValue structure in order to easily take advantage of float,int,uint borderColor types.

2) For hardware which supports a limited number of border colors what happens if that number is exceeded? Should this be handled by the driver unbeknownst to the application? In Revision 1 we had solved this issue using a new Object type, however that may have lead to additional system resource consumption which would otherwise not be required.

RESOLVED: Added PhysicalDeviceCustomBorderColorPropertiesEXT::maxCustomBorderColorSamplers for tracking implementation-specific limit, and Valid Usage statement handling overflow.

3) Should this be supported for immutable samplers at all, or by a feature bit? Some implementations may not be able to support custom border colors on immutable samplers — is it worthwhile enabling this to work on them for implementations that can support it, or forbidding it entirely.

RESOLVED: Samplers created with a custom border color are forbidden from being immutable. This resolves concerns for implementations where the custom border color is an index to a LUT instead of being directly embedded into sampler state.

4) Should UINT and SINT (unsigned integer and signed integer) border color types be separated or should they be combined into one generic INT (integer) type?

RESOLVED: Separating these doesn’t make much sense as the existing fixed border color types don’t have this distinction, and there is no reason in hardware to do so. This separation would also create unnecessary work and considerations for the application.

Version History

Revision 1, 2019-10-10 (Joshua Ashton)

Internal revisions.

Revision 2, 2019-10-11 (Liam Middlebrook)

Remove VkCustomBorderColor object and associated functions

Add issues concerning HW limitations for custom border color count

Revision 3, 2019-10-12 (Joshua Ashton)

Re-expose the limits for the maximum number of unique border colors

Add extra details about border color tracking

Fix typos

Revision 4, 2019-10-12 (Joshua Ashton)

Changed maxUniqueCustomBorderColors to a uint32_t from a VkDeviceSize

Revision 5, 2019-10-14 (Liam Middlebrook)

Added features bit

Revision 6, 2019-10-15 (Joshua Ashton)

Type-ize VK_BORDER_COLOR_CUSTOM

Fix const-ness on pNext of VkSamplerCustomBorderColorCreateInfoEXT

Revision 7, 2019-11-26 (Liam Middlebrook)

Renamed maxUniqueCustomBorderColors to maxCustomBorderColors

Revision 8, 2019-11-29 (Joshua Ashton)

Renamed borderColor member of VkSamplerCustomBorderColorCreateInfoEXT to customBorderColor

Revision 9, 2020-02-19 (Joshua Ashton)

Renamed maxCustomBorderColors to maxCustomBorderColorSamplers

Revision 10, 2020-02-21 (Joshua Ashton)

Added format to VkSamplerCustomBorderColorCreateInfoEXT and feature bit

Revision 11, 2020-04-07 (Joshua Ashton)

Dropped UINT/SINT border color differences, consolidated types

Revision 12, 2020-04-16 (Joshua Ashton)

Renamed VK_BORDER_COLOR_CUSTOM_FLOAT_EXT to VK_BORDER_COLOR_FLOAT_CUSTOM_EXT for consistency

See Also

PhysicalDeviceCustomBorderColorFeaturesEXT, PhysicalDeviceCustomBorderColorPropertiesEXT, SamplerCustomBorderColorCreateInfoEXT

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data SamplerCustomBorderColorCreateInfoEXT

VkSamplerCustomBorderColorCreateInfoEXT - Structure specifying custom border color

Valid Usage

If provided format is not FORMAT_UNDEFINED then the SamplerCreateInfo::borderColor type must match the sampled type of the provided format, as shown in the SPIR-V Sampled Type column of the https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#formats-numericformat table

If the customBorderColorWithoutFormat feature is not enabled then format must not be FORMAT_UNDEFINED

If the sampler is used to sample an image view of FORMAT_B4G4R4A4_UNORM_PACK16, FORMAT_B5G6R5_UNORM_PACK16, or FORMAT_B5G5R5A1_UNORM_PACK16 format then format must not be FORMAT_UNDEFINED

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_SAMPLER_CUSTOM_BORDER_COLOR_CREATE_INFO_EXT

format must be a valid Format value

See Also

ClearColorValue, Format, StructureType

Constructors

SamplerCustomBorderColorCreateInfoEXT

Fields

customBorderColor :: ClearColorValue

customBorderColor is a ClearColorValue representing the desired custom sampler border color.

format :: Format

format is a Format representing the format of the sampled image view(s). This field may be FORMAT_UNDEFINED if the customBorderColorWithoutFormat feature is enabled.

data PhysicalDeviceCustomBorderColorPropertiesEXT

VkPhysicalDeviceCustomBorderColorPropertiesEXT - Structure describing whether custom border colors can be supported by an implementation

Members

The members of the PhysicalDeviceCustomBorderColorPropertiesEXT structure describe the following features:

Valid Usage (Implicit)

See Also

StructureType

Constructors

PhysicalDeviceCustomBorderColorPropertiesEXT

Fields

maxCustomBorderColorSamplers :: Word32

maxCustomBorderColorSamplers indicates the maximum number of samplers with custom border colors which can simultaneously exist on a device.

data PhysicalDeviceCustomBorderColorFeaturesEXT

VkPhysicalDeviceCustomBorderColorFeaturesEXT - Structure describing whether custom border colors can be supported by an implementation

Members

The members of the PhysicalDeviceCustomBorderColorFeaturesEXT structure describe the following features:

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceCustomBorderColorFeaturesEXT

Fields

customBorderColors :: Bool

customBorderColors indicates that the implementation supports providing a borderColor value with one of the following values at sampler creation time:

BORDER_COLOR_FLOAT_CUSTOM_EXT

BORDER_COLOR_INT_CUSTOM_EXT

customBorderColorWithoutFormat :: Bool

customBorderColorWithoutFormat indicates that explicit formats are not required for custom border colors and the value of the format member of the SamplerCustomBorderColorCreateInfoEXT structure may be FORMAT_UNDEFINED. If this feature bit is not set, applications must provide the Format of the image view(s) being sampled by this sampler in the format member of the SamplerCustomBorderColorCreateInfoEXT structure.

type EXT_CUSTOM_BORDER_COLOR_SPEC_VERSION = 12

pattern EXT_CUSTOM_BORDER_COLOR_SPEC_VERSION :: forall a. Integral a => a

type EXT_CUSTOM_BORDER_COLOR_EXTENSION_NAME = "VK_EXT_custom_border_color"

pattern EXT_CUSTOM_BORDER_COLOR_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a