Vulkan.Extensions.VK_EXT_debug_marker

Documentation

debugMarkerSetObjectNameEXT

Arguments

:: forall io. MonadIO io

=> Device
device is the device that created the object.

-> DebugMarkerObjectNameInfoEXT
pNameInfo is a pointer to a DebugMarkerObjectNameInfoEXT structure specifying the parameters of the name to set on the object.

-> io ()

vkDebugMarkerSetObjectNameEXT - Give a user-friendly name to an object

Valid Usage (Implicit)

device must be a valid Device handle

pNameInfo must be a valid pointer to a valid DebugMarkerObjectNameInfoEXT structure

Host Synchronization

Host access to pNameInfo->object must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

DebugMarkerObjectNameInfoEXT, Device

debugMarkerSetObjectTagEXT

Arguments

:: forall io. MonadIO io

=> Device
device is the device that created the object.

-> DebugMarkerObjectTagInfoEXT
pTagInfo is a pointer to a DebugMarkerObjectTagInfoEXT structure specifying the parameters of the tag to attach to the object.

-> io ()

vkDebugMarkerSetObjectTagEXT - Attach arbitrary data to an object

Valid Usage (Implicit)

device must be a valid Device handle

pTagInfo must be a valid pointer to a valid DebugMarkerObjectTagInfoEXT structure

Host Synchronization

Host access to pTagInfo->object must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

DebugMarkerObjectTagInfoEXT, Device

cmdDebugMarkerBeginEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command is recorded.

-> DebugMarkerMarkerInfoEXT
pMarkerInfo is a pointer to a DebugMarkerMarkerInfoEXT structure specifying the parameters of the marker region to open.

-> io ()

vkCmdDebugMarkerBeginEXT - Open a command buffer marker region

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pMarkerInfo must be a valid pointer to a valid DebugMarkerMarkerInfoEXT structure

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute

See Also

CommandBuffer, DebugMarkerMarkerInfoEXT

cmdDebugMarkerEndEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command is recorded.

-> io ()

vkCmdDebugMarkerEndEXT - Close a command buffer marker region

Description

An application may open a marker region in one command buffer and close it in another, or otherwise split marker regions across multiple command buffers or multiple queue submissions. When viewed from the linear series of submissions to a single queue, the calls to cmdDebugMarkerBeginEXT and cmdDebugMarkerEndEXT must be matched and balanced.

Valid Usage

There must be an outstanding cmdDebugMarkerBeginEXT command prior to the cmdDebugMarkerEndEXT on the queue that commandBuffer is submitted to

If commandBuffer is a secondary command buffer, there must be an outstanding cmdDebugMarkerBeginEXT command recorded to commandBuffer that has not previously been ended by a call to cmdDebugMarkerEndEXT

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute

See Also

CommandBuffer

cmdDebugMarkerInsertEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command is recorded.

-> DebugMarkerMarkerInfoEXT
pMarkerInfo is a pointer to a DebugMarkerMarkerInfoEXT structure specifying the parameters of the marker to insert.

-> io ()

vkCmdDebugMarkerInsertEXT - Insert a marker label into a command buffer

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pMarkerInfo must be a valid pointer to a valid DebugMarkerMarkerInfoEXT structure

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute

See Also

CommandBuffer, DebugMarkerMarkerInfoEXT

data DebugMarkerObjectNameInfoEXT

VkDebugMarkerObjectNameInfoEXT - Specify parameters of a name to give to an object

Description

Applications may change the name associated with an object simply by calling debugMarkerSetObjectNameEXT again with a new string. To remove a previously set name, pObjectName should be set to an empty string.

Valid Usage (Implicit)

See Also

DebugReportObjectTypeEXT, StructureType, debugMarkerSetObjectNameEXT

Constructors

DebugMarkerObjectNameInfoEXT

Fields

objectType :: DebugReportObjectTypeEXT

objectType is a DebugReportObjectTypeEXT specifying the type of the object to be named.

objectType must not be DEBUG_REPORT_OBJECT_TYPE_UNKNOWN_EXT

objectType must be a valid DebugReportObjectTypeEXT value

object :: Word64

object is the object to be named.

object must not be NULL_HANDLE

object must be a Vulkan object of the type associated with objectType as defined in https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#debug-report-object-types

objectName :: ByteString

pObjectName is a null-terminated UTF-8 string specifying the name to apply to object.

pObjectName must be a null-terminated UTF-8 string

data DebugMarkerObjectTagInfoEXT

VkDebugMarkerObjectTagInfoEXT - Specify parameters of a tag to attach to an object

Description

The tagName parameter gives a name or identifier to the type of data being tagged. This can be used by debugging layers to easily filter for only data that can be used by that implementation.

Valid Usage (Implicit)

See Also

DebugReportObjectTypeEXT, StructureType, debugMarkerSetObjectTagEXT

Constructors

DebugMarkerObjectTagInfoEXT

Fields

objectType :: DebugReportObjectTypeEXT

objectType is a DebugReportObjectTypeEXT specifying the type of the object to be named.

objectType must not be DEBUG_REPORT_OBJECT_TYPE_UNKNOWN_EXT

objectType must be a valid DebugReportObjectTypeEXT value

object :: Word64

object is the object to be tagged.

object must not be NULL_HANDLE

object must be a Vulkan object of the type associated with objectType as defined in https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#debug-report-object-types

tagName :: Word64

tagName is a numerical identifier of the tag.

tagSize :: Word64

tagSize is the number of bytes of data to attach to the object.

tagSize must be greater than 0

tag :: Ptr ()

pTag is a pointer to an array of tagSize bytes containing the data to be associated with the object.

pTag must be a valid pointer to an array of tagSize bytes

data DebugMarkerMarkerInfoEXT

VkDebugMarkerMarkerInfoEXT - Specify parameters of a command buffer marker region

Valid Usage (Implicit)

See Also

StructureType, cmdDebugMarkerBeginEXT, cmdDebugMarkerInsertEXT

Constructors

DebugMarkerMarkerInfoEXT

Fields

markerName :: ByteString

pMarkerName is a pointer to a null-terminated UTF-8 string containing the name of the marker.

pMarkerName must be a null-terminated UTF-8 string

color :: (Float, Float, Float, Float)

color is an optional RGBA color value that can be associated with the marker. A particular implementation may choose to ignore this color value. The values contain RGBA values in order, in the range 0.0 to 1.0. If all elements in color are set to 0.0 then it is ignored.

type EXT_DEBUG_MARKER_SPEC_VERSION = 4

pattern EXT_DEBUG_MARKER_SPEC_VERSION :: forall a. Integral a => a

type EXT_DEBUG_MARKER_EXTENSION_NAME = "VK_EXT_debug_marker"

pattern EXT_DEBUG_MARKER_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype DebugReportObjectTypeEXT

VkDebugReportObjectTypeEXT - Specify the type of an object handle

Description

'

DebugReportObjectTypeEXT 	 Vulkan Handle Type

DEBUG_REPORT_OBJECT_TYPE_UNKNOWN_EXT 	 Unknown/Undefined Handle

DEBUG_REPORT_OBJECT_TYPE_INSTANCE_EXT 	 Instance

DEBUG_REPORT_OBJECT_TYPE_PHYSICAL_DEVICE_EXT 	 PhysicalDevice

DEBUG_REPORT_OBJECT_TYPE_DEVICE_EXT 	 Device

DEBUG_REPORT_OBJECT_TYPE_QUEUE_EXT 	 Queue

DEBUG_REPORT_OBJECT_TYPE_SEMAPHORE_EXT 	 Semaphore

DEBUG_REPORT_OBJECT_TYPE_COMMAND_BUFFER_EXT 	 CommandBuffer

DEBUG_REPORT_OBJECT_TYPE_FENCE_EXT 	 Fence

DEBUG_REPORT_OBJECT_TYPE_DEVICE_MEMORY_EXT 	 DeviceMemory

DEBUG_REPORT_OBJECT_TYPE_BUFFER_EXT 	 Buffer

DEBUG_REPORT_OBJECT_TYPE_IMAGE_EXT 	 Image

DEBUG_REPORT_OBJECT_TYPE_EVENT_EXT 	 Event

DEBUG_REPORT_OBJECT_TYPE_QUERY_POOL_EXT 	 QueryPool

DEBUG_REPORT_OBJECT_TYPE_BUFFER_VIEW_EXT 	 BufferView

DEBUG_REPORT_OBJECT_TYPE_IMAGE_VIEW_EXT 	 ImageView

DEBUG_REPORT_OBJECT_TYPE_SHADER_MODULE_EXT 	 ShaderModule

DEBUG_REPORT_OBJECT_TYPE_PIPELINE_CACHE_EXT 	 PipelineCache

DEBUG_REPORT_OBJECT_TYPE_PIPELINE_LAYOUT_EXT 	 PipelineLayout

DEBUG_REPORT_OBJECT_TYPE_RENDER_PASS_EXT 	 RenderPass

DEBUG_REPORT_OBJECT_TYPE_PIPELINE_EXT 	 Pipeline

DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_LAYOUT_EXT 	 DescriptorSetLayout

DEBUG_REPORT_OBJECT_TYPE_SAMPLER_EXT 	 Sampler

DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_POOL_EXT 	 DescriptorPool

DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_EXT 	 DescriptorSet

DEBUG_REPORT_OBJECT_TYPE_FRAMEBUFFER_EXT 	 Framebuffer

DEBUG_REPORT_OBJECT_TYPE_COMMAND_POOL_EXT 	 CommandPool

DEBUG_REPORT_OBJECT_TYPE_SURFACE_KHR_EXT 	 SurfaceKHR

DEBUG_REPORT_OBJECT_TYPE_SWAPCHAIN_KHR_EXT 	 SwapchainKHR

DEBUG_REPORT_OBJECT_TYPE_DEBUG_REPORT_CALLBACK_EXT_EXT 	 DebugReportCallbackEXT

DEBUG_REPORT_OBJECT_TYPE_DISPLAY_KHR_EXT 	 DisplayKHR

DEBUG_REPORT_OBJECT_TYPE_DISPLAY_MODE_KHR_EXT 	 DisplayModeKHR

DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_EXT 	 DescriptorUpdateTemplate

DebugReportObjectTypeEXT and Vulkan Handle Relationship

Note

The primary expected use of ERROR_VALIDATION_FAILED_EXT is for validation layer testing. It is not expected that an application would see this error code during normal use of the validation layers.

See Also

DebugMarkerObjectNameInfoEXT, DebugMarkerObjectTagInfoEXT, debugReportMessageEXT

Constructors

DebugReportObjectTypeEXT Int32

Bundled Patterns

pattern DEBUG_REPORT_OBJECT_TYPE_UNKNOWN_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_INSTANCE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_PHYSICAL_DEVICE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DEVICE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_QUEUE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SEMAPHORE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_COMMAND_BUFFER_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_FENCE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DEVICE_MEMORY_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_BUFFER_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_IMAGE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_EVENT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_QUERY_POOL_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_BUFFER_VIEW_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_IMAGE_VIEW_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SHADER_MODULE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_PIPELINE_CACHE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_PIPELINE_LAYOUT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_RENDER_PASS_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_PIPELINE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_LAYOUT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SAMPLER_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_POOL_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_FRAMEBUFFER_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_COMMAND_POOL_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SURFACE_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SWAPCHAIN_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DEBUG_REPORT_CALLBACK_EXT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DISPLAY_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DISPLAY_MODE_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_VALIDATION_CACHE_EXT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_ACCELERATION_STRUCTURE_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_EXT :: DebugReportObjectTypeEXT