Vulkan.Extensions.VK_EXT_depth_clip_enable

Description

Name

VK_EXT_depth_clip_enable - device extension

VK_EXT_depth_clip_enable

Name String

VK_EXT_depth_clip_enable

Extension Type

Device extension

Registered Extension Number

103

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Special Use

D3D support

Contact

Piers Daniell https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_EXT_depth_clip_enable:%20&body=@pdaniell-nv%20

Other Extension Metadata

Last Modified Date

2018-12-20

Contributors

Daniel Rakos, AMD

Henri Verbeet, CodeWeavers

Jeff Bolz, NVIDIA

Philip Rebohle, DXVK

Tobias Hector, AMD

Description

This extension allows the depth clipping operation, that is normally implicitly controlled by PipelineRasterizationStateCreateInfo::depthClampEnable, to instead be controlled explicitly by PipelineRasterizationDepthClipStateCreateInfoEXT::depthClipEnable.

This is useful for translating DX content which assumes depth clamping is always enabled, but depth clip can be controlled by the DepthClipEnable rasterization state (D3D12_RASTERIZER_DESC).

New Structures

Extending PhysicalDeviceFeatures2, DeviceCreateInfo:

PhysicalDeviceDepthClipEnableFeaturesEXT

Extending PipelineRasterizationStateCreateInfo:

PipelineRasterizationDepthClipStateCreateInfoEXT

New Bitmasks

PipelineRasterizationDepthClipStateCreateFlagsEXT

New Enum Constants

EXT_DEPTH_CLIP_ENABLE_EXTENSION_NAME

EXT_DEPTH_CLIP_ENABLE_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_PHYSICAL_DEVICE_DEPTH_CLIP_ENABLE_FEATURES_EXT

STRUCTURE_TYPE_PIPELINE_RASTERIZATION_DEPTH_CLIP_STATE_CREATE_INFO_EXT

Version History

Revision 1, 2018-12-20 (Piers Daniell)

Internal revisions

See Also

PhysicalDeviceDepthClipEnableFeaturesEXT, PipelineRasterizationDepthClipStateCreateFlagsEXT, PipelineRasterizationDepthClipStateCreateInfoEXT

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data PhysicalDeviceDepthClipEnableFeaturesEXT

VkPhysicalDeviceDepthClipEnableFeaturesEXT - Structure indicating support for explicit enable of depth clip

Members

The members of the PhysicalDeviceDepthClipEnableFeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceDepthClipEnableFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether the feature is supported. PhysicalDeviceDepthClipEnableFeaturesEXT can also be included in the pNext chain of DeviceCreateInfo to enable this feature.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceDepthClipEnableFeaturesEXT

Fields

depthClipEnable :: Bool

depthClipEnable indicates that the implementation supports setting the depth clipping operation explicitly via the PipelineRasterizationDepthClipStateCreateInfoEXT pipeline state. Otherwise depth clipping is only enabled when PipelineRasterizationStateCreateInfo::depthClampEnable is set to FALSE.

data PipelineRasterizationDepthClipStateCreateInfoEXT

VkPipelineRasterizationDepthClipStateCreateInfoEXT - Structure specifying depth clipping state

Valid Usage (Implicit)

See Also

Bool32, PipelineRasterizationDepthClipStateCreateFlagsEXT, StructureType

Constructors

PipelineRasterizationDepthClipStateCreateInfoEXT

Fields

flags :: PipelineRasterizationDepthClipStateCreateFlagsEXT

flags is reserved for future use.

flags must be 0

depthClipEnable :: Bool

depthClipEnable controls whether depth clipping is enabled as described in Primitive Clipping.

newtype PipelineRasterizationDepthClipStateCreateFlagsEXT

VkPipelineRasterizationDepthClipStateCreateFlagsEXT - Reserved for future use

Description

PipelineRasterizationDepthClipStateCreateFlagsEXT is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

PipelineRasterizationDepthClipStateCreateInfoEXT

Constructors

PipelineRasterizationDepthClipStateCreateFlagsEXT Flags

type EXT_DEPTH_CLIP_ENABLE_SPEC_VERSION = 1

pattern EXT_DEPTH_CLIP_ENABLE_SPEC_VERSION :: forall a. Integral a => a

type EXT_DEPTH_CLIP_ENABLE_EXTENSION_NAME = "VK_EXT_depth_clip_enable"

pattern EXT_DEPTH_CLIP_ENABLE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a