Vulkan.Extensions.VK_EXT_depth_range_unrestricted

Description

Name

VK_EXT_depth_range_unrestricted - device extension

VK_EXT_depth_range_unrestricted

Name String

VK_EXT_depth_range_unrestricted

Extension Type

Device extension

Registered Extension Number

14

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Contact

Piers Daniell https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_EXT_depth_range_unrestricted:%20&body=@pdaniell-nv%20

Other Extension Metadata

Last Modified Date

2017-06-22

Contributors

Daniel Koch, NVIDIA

Jeff Bolz, NVIDIA

Description

This extension removes the Viewport minDepth and maxDepth restrictions that the values must be between 0.0 and 1.0, inclusive. It also removes the same restriction on PipelineDepthStencilStateCreateInfo minDepthBounds and maxDepthBounds. Finally it removes the restriction on the depth value in ClearDepthStencilValue.

New Enum Constants

EXT_DEPTH_RANGE_UNRESTRICTED_EXTENSION_NAME

EXT_DEPTH_RANGE_UNRESTRICTED_SPEC_VERSION

Issues

1) How do Viewport minDepth and maxDepth values outside of the 0.0 to 1.0 range interact with Primitive Clipping?

RESOLVED: The behavior described in Primitive Clipping still applies. If depth clamping is disabled the depth values are still clipped to 0 ≤ zc ≤ wc before the viewport transform. If depth clamping is enabled the above equation is ignored and the depth values are instead clamped to the Viewport minDepth and maxDepth values, which in the case of this extension can be outside of the 0.0 to 1.0 range.

2) What happens if a resulting depth fragment is outside of the 0.0 to 1.0 range and the depth buffer is fixed-point rather than floating-point?

RESOLVED: The supported range of a fixed-point depth buffer is 0.0 to 1.0 and depth fragments are clamped to this range.

Version History

Revision 1, 2017-06-22 (Piers Daniell)

Internal revisions

See Also

No cross-references are available

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

type EXT_DEPTH_RANGE_UNRESTRICTED_SPEC_VERSION = 1

pattern EXT_DEPTH_RANGE_UNRESTRICTED_SPEC_VERSION :: forall a. Integral a => a

type EXT_DEPTH_RANGE_UNRESTRICTED_EXTENSION_NAME = "VK_EXT_depth_range_unrestricted"

pattern EXT_DEPTH_RANGE_UNRESTRICTED_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a