Vulkan.Extensions.VK_EXT_direct_mode_display

Description

Name

VK_EXT_direct_mode_display - instance extension

VK_EXT_direct_mode_display

Name String

VK_EXT_direct_mode_display

Extension Type

Instance extension

Registered Extension Number

89

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_display

Contact

James Jones https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_EXT_direct_mode_display:%20&body=@cubanismo%20

Other Extension Metadata

Last Modified Date

2016-12-13

IP Status

No known IP claims.

Contributors

Pierre Boudier, NVIDIA

James Jones, NVIDIA

Damien Leone, NVIDIA

Pierre-Loup Griffais, Valve

Liam Middlebrook, NVIDIA

Description

This is extension, along with related platform extensions, allows applications to take exclusive control of displays associated with a native windowing system. This is especially useful for virtual reality applications that wish to hide HMDs (head mounted displays) from the native platform’s display management system, desktop, and/or other applications.

New Commands

releaseDisplayEXT

New Enum Constants

EXT_DIRECT_MODE_DISPLAY_EXTENSION_NAME

EXT_DIRECT_MODE_DISPLAY_SPEC_VERSION

Issues

1) Should this extension and its related platform-specific extensions leverage VK_KHR_display, or provide separate equivalent interfaces.

RESOLVED: Use VK_KHR_display concepts and objects. VK_KHR_display can be used to enumerate all displays on the system, including those attached to/in use by a window system or native platform, but VK_KHR_display_swapchain will fail to create a swapchain on in-use displays. This extension and its platform-specific children will allow applications to grab in-use displays away from window systems and/or native platforms, allowing them to be used with VK_KHR_display_swapchain.

2) Are separate calls needed to acquire displays and enable direct mode?

RESOLVED: No, these operations happen in one combined command. Acquiring a display puts it into direct mode.

Version History

Revision 1, 2016-12-13 (James Jones)

Initial draft

See Also

releaseDisplayEXT

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

releaseDisplayEXT

Arguments

:: forall io. MonadIO io

=> PhysicalDevice
physicalDevice The physical device the display is on.

physicalDevice must be a valid PhysicalDevice handle

-> DisplayKHR
display The display to release control of.

display must be a valid DisplayKHR handle

display must have been created, allocated, or retrieved from physicalDevice

-> io ()

vkReleaseDisplayEXT - Release access to an acquired VkDisplayKHR

Return Codes

Success

SUCCESS

See Also

DisplayKHR, PhysicalDevice

type EXT_DIRECT_MODE_DISPLAY_SPEC_VERSION = 1

pattern EXT_DIRECT_MODE_DISPLAY_SPEC_VERSION :: forall a. Integral a => a

type EXT_DIRECT_MODE_DISPLAY_EXTENSION_NAME = "VK_EXT_direct_mode_display"

pattern EXT_DIRECT_MODE_DISPLAY_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype DisplayKHR

VkDisplayKHR - Opaque handle to a display object

See Also

DisplayPlanePropertiesKHR, DisplayPropertiesKHR, acquireXlibDisplayEXT, createDisplayModeKHR, displayPowerControlEXT, getDisplayModeProperties2KHR, getDisplayModePropertiesKHR, getDisplayPlaneSupportedDisplaysKHR, getRandROutputDisplayEXT, registerDisplayEventEXT, releaseDisplayEXT

Constructors

DisplayKHR Word64