Vulkan.Extensions.VK_EXT_display_surface_counter

Documentation

getPhysicalDeviceSurfaceCapabilities2EXT

Arguments

:: forall io. MonadIO io

=> PhysicalDevice
physicalDevice is the physical device that will be associated with the swapchain to be created, as described for createSwapchainKHR.

-> SurfaceKHR
surface is the surface that will be associated with the swapchain.

-> io SurfaceCapabilities2EXT

vkGetPhysicalDeviceSurfaceCapabilities2EXT - Query surface capabilities

Description

getPhysicalDeviceSurfaceCapabilities2EXT behaves similarly to getPhysicalDeviceSurfaceCapabilitiesKHR, with the ability to return extended information by adding extending structures to the pNext chain of its pSurfaceCapabilities parameter.

Valid Usage (Implicit)

physicalDevice must be a valid PhysicalDevice handle

surface must be a valid SurfaceKHR handle

pSurfaceCapabilities must be a valid pointer to a SurfaceCapabilities2EXT structure

Both of physicalDevice, and surface must have been created, allocated, or retrieved from the same Instance

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_SURFACE_LOST_KHR

See Also

PhysicalDevice, SurfaceCapabilities2EXT, SurfaceKHR

pattern STRUCTURE_TYPE_SURFACE_CAPABILITIES2_EXT :: StructureType

data SurfaceCapabilities2EXT

VkSurfaceCapabilities2EXT - Structure describing capabilities of a surface

Members

All members of SurfaceCapabilities2EXT are identical to the corresponding members of SurfaceCapabilitiesKHR where one exists. The remaining members are:

Valid Usage (Implicit)

See Also

CompositeAlphaFlagsKHR, Extent2D, ImageUsageFlags, StructureType, SurfaceCounterFlagsEXT, SurfaceTransformFlagBitsKHR, SurfaceTransformFlagsKHR, getPhysicalDeviceSurfaceCapabilities2EXT

Constructors

SurfaceCapabilities2EXT

Fields

minImageCount :: Word32

maxImageCount :: Word32

currentExtent :: Extent2D

minImageExtent :: Extent2D

maxImageExtent :: Extent2D

maxImageArrayLayers :: Word32

supportedTransforms :: SurfaceTransformFlagsKHR

currentTransform :: SurfaceTransformFlagBitsKHR

supportedCompositeAlpha :: CompositeAlphaFlagsKHR

supportedUsageFlags :: ImageUsageFlags

supportedSurfaceCounters :: SurfaceCounterFlagsEXT

supportedSurfaceCounters is a bitmask of SurfaceCounterFlagBitsEXT indicating the supported surface counter types.

supportedSurfaceCounters must not include SURFACE_COUNTER_VBLANK_EXT unless the surface queried is a display surface

newtype SurfaceCounterFlagBitsEXT

VkSurfaceCounterFlagBitsEXT - Surface-relative counter types

See Also

SurfaceCounterFlagsEXT, getSwapchainCounterEXT

Constructors

SurfaceCounterFlagBitsEXT Flags

Bundled Patterns

pattern SURFACE_COUNTER_VBLANK_EXT :: SurfaceCounterFlagBitsEXT
SURFACE_COUNTER_VBLANK_EXT specifies a counter incrementing once every time a vertical blanking period occurs on the display associated with the surface.

type SurfaceCounterFlagsEXT = SurfaceCounterFlagBitsEXT

type EXT_DISPLAY_SURFACE_COUNTER_SPEC_VERSION = 1

pattern EXT_DISPLAY_SURFACE_COUNTER_SPEC_VERSION :: forall a. Integral a => a

type EXT_DISPLAY_SURFACE_COUNTER_EXTENSION_NAME = "VK_EXT_display_surface_counter"

pattern EXT_DISPLAY_SURFACE_COUNTER_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype SurfaceKHR

VkSurfaceKHR - Opaque handle to a surface object

Description

The VK_KHR_surface extension declares the SurfaceKHR object, and provides a function for destroying SurfaceKHR objects. Separate platform-specific extensions each provide a function for creating a SurfaceKHR object for the respective platform. From the application’s perspective this is an opaque handle, just like the handles of other Vulkan objects.

See Also

PhysicalDeviceSurfaceInfo2KHR, SwapchainCreateInfoKHR, createAndroidSurfaceKHR, createDirectFBSurfaceEXT, createDisplayPlaneSurfaceKHR, createHeadlessSurfaceEXT, createIOSSurfaceMVK, createImagePipeSurfaceFUCHSIA, createMacOSSurfaceMVK, createMetalSurfaceEXT, createStreamDescriptorSurfaceGGP, createViSurfaceNN, createWaylandSurfaceKHR, createWin32SurfaceKHR, createXcbSurfaceKHR, createXlibSurfaceKHR, destroySurfaceKHR, getDeviceGroupSurfacePresentModesKHR, getPhysicalDevicePresentRectanglesKHR, getPhysicalDeviceSurfaceCapabilities2EXT, getPhysicalDeviceSurfaceCapabilitiesKHR, getPhysicalDeviceSurfaceFormatsKHR, getPhysicalDeviceSurfacePresentModesKHR, getPhysicalDeviceSurfaceSupportKHR

Constructors

SurfaceKHR Word64

newtype CompositeAlphaFlagBitsKHR

VkCompositeAlphaFlagBitsKHR - alpha compositing modes supported on a device

Description

These values are described as follows:

See Also

CompositeAlphaFlagsKHR, SwapchainCreateInfoKHR

Constructors

CompositeAlphaFlagBitsKHR Flags

Bundled Patterns

pattern COMPOSITE_ALPHA_OPAQUE_BIT_KHR :: CompositeAlphaFlagBitsKHR
COMPOSITE_ALPHA_OPAQUE_BIT_KHR: The alpha channel, if it exists, of the images is ignored in the compositing process. Instead, the image is treated as if it has a constant alpha of 1.0.

pattern COMPOSITE_ALPHA_PRE_MULTIPLIED_BIT_KHR :: CompositeAlphaFlagBitsKHR
COMPOSITE_ALPHA_PRE_MULTIPLIED_BIT_KHR: The alpha channel, if it exists, of the images is respected in the compositing process. The non-alpha channels of the image are expected to already be multiplied by the alpha channel by the application.

pattern COMPOSITE_ALPHA_POST_MULTIPLIED_BIT_KHR :: CompositeAlphaFlagBitsKHR
COMPOSITE_ALPHA_POST_MULTIPLIED_BIT_KHR: The alpha channel, if it exists, of the images is respected in the compositing process. The non-alpha channels of the image are not expected to already be multiplied by the alpha channel by the application; instead, the compositor will multiply the non-alpha channels of the image by the alpha channel during compositing.

pattern COMPOSITE_ALPHA_INHERIT_BIT_KHR :: CompositeAlphaFlagBitsKHR
COMPOSITE_ALPHA_INHERIT_BIT_KHR: The way in which the presentation engine treats the alpha channel in the images is unknown to the Vulkan API. Instead, the application is responsible for setting the composite alpha blending mode using native window system commands. If the application does not set the blending mode using native window system commands, then a platform-specific default will be used.

type CompositeAlphaFlagsKHR = CompositeAlphaFlagBitsKHR

newtype SurfaceTransformFlagBitsKHR

VkSurfaceTransformFlagBitsKHR - presentation transforms supported on a device

See Also

CommandBufferInheritanceRenderPassTransformInfoQCOM, DisplaySurfaceCreateInfoKHR, RenderPassTransformBeginInfoQCOM, SurfaceCapabilities2EXT, SurfaceCapabilitiesKHR, SurfaceTransformFlagsKHR, SwapchainCreateInfoKHR

Constructors

SurfaceTransformFlagBitsKHR Flags

Bundled Patterns

pattern SURFACE_TRANSFORM_IDENTITY_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_IDENTITY_BIT_KHR specifies that image content is presented without being transformed.

pattern SURFACE_TRANSFORM_ROTATE_90_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_ROTATE_90_BIT_KHR specifies that image content is rotated 90 degrees clockwise.

pattern SURFACE_TRANSFORM_ROTATE_180_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_ROTATE_180_BIT_KHR specifies that image content is rotated 180 degrees clockwise.

pattern SURFACE_TRANSFORM_ROTATE_270_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_ROTATE_270_BIT_KHR specifies that image content is rotated 270 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_BIT_KHR specifies that image content is mirrored horizontally.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_90_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_90_BIT_KHR specifies that image content is mirrored horizontally, then rotated 90 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_180_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_180_BIT_KHR specifies that image content is mirrored horizontally, then rotated 180 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_270_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_270_BIT_KHR specifies that image content is mirrored horizontally, then rotated 270 degrees clockwise.

pattern SURFACE_TRANSFORM_INHERIT_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_INHERIT_BIT_KHR specifies that the presentation transform is not specified, and is instead determined by platform-specific considerations and mechanisms outside Vulkan.

type SurfaceTransformFlagsKHR = SurfaceTransformFlagBitsKHR