Vulkan.Extensions.VK_EXT_extended_dynamic_state

Documentation

cmdSetCullModeEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command will be recorded.

-> CullModeFlags
cullMode specifies the cull mode property to use for drawing.

-> io ()

vkCmdSetCullModeEXT - Set the cull mode property

Valid Usage

The extendedDynamicState feature must be enabled

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

cullMode must be a valid combination of CullModeFlagBits values

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

CommandBuffer, CullModeFlags

cmdSetFrontFaceEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command will be recorded.

-> FrontFace
frontFace specifies the front face property to use for drawing.

-> io ()

vkCmdSetFrontFaceEXT - Set the front face property

Valid Usage

The extendedDynamicState feature must be enabled

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

frontFace must be a valid FrontFace value

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

CommandBuffer, FrontFace

cmdSetPrimitiveTopologyEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command will be recorded.

-> PrimitiveTopology

-> io ()

vkCmdSetPrimitiveTopologyEXT - Set the primitive topology state

Valid Usage

The extendedDynamicState feature must be enabled

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

primitiveTopology must be a valid PrimitiveTopology value

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

CommandBuffer, PrimitiveTopology

cmdSetViewportWithCountEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command will be recorded.

-> ("viewports" ::: Vector Viewport)
pViewports specifies the viewports to use for drawing.

-> io ()

vkCmdSetViewportWithCountEXT - Set the viewport count and viewports

Valid Usage

The extendedDynamicState feature must be enabled

viewportCount must be between 1 and PhysicalDeviceLimits::maxViewports, inclusive

If the multiple viewports feature is not enabled, viewportCount must be 1

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pViewports must be a valid pointer to an array of viewportCount valid Viewport structures

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

viewportCount must be greater than 0

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

CommandBuffer, Viewport

cmdSetScissorWithCountEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command will be recorded.

-> ("scissors" ::: Vector Rect2D)
pScissors specifies the scissors to use for drawing.

-> io ()

vkCmdSetScissorWithCountEXT - Set the scissor count and scissors

Valid Usage

The extendedDynamicState feature must be enabled

scissorCount must be between 1 and PhysicalDeviceLimits::maxViewports, inclusive

If the multiple viewports feature is not enabled, scissorCount must be 1

The x and y members of offset member of any element of pScissors must be greater than or equal to 0

Evaluation of (offset.x + extent.width) must not cause a signed integer addition overflow for any element of pScissors

Evaluation of (offset.y + extent.height) must not cause a signed integer addition overflow for any element of pScissors

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pScissors must be a valid pointer to an array of scissorCount Rect2D structures

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

scissorCount must be greater than 0

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

CommandBuffer, Rect2D

cmdBindVertexBuffers2EXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command is recorded.

-> ("firstBinding" ::: Word32)
firstBinding is the index of the first vertex input binding whose state is updated by the command.

-> ("buffers" ::: Vector Buffer)
pBuffers is a pointer to an array of buffer handles.

-> ("offsets" ::: Vector DeviceSize)
pOffsets is a pointer to an array of buffer offsets.

-> ("sizes" ::: Vector DeviceSize)
pSizes is an optional array of the size in bytes of vertex data bound from pBuffers.

-> ("strides" ::: Vector DeviceSize)
pStrides is optional, and when not NULL is a pointer to an array of buffer strides.

-> io ()

vkCmdBindVertexBuffers2EXT - Bind vertex buffers to a command buffer

Description

The values taken from elements i of pBuffers and pOffsets replace the current state for the vertex input binding firstBinding + i, for i in [0, bindingCount). The vertex input binding is updated to start at the offset indicated by pOffsets[i] from the start of the buffer pBuffers[i]. If pSizes is not NULL then pSizes[i] specifies the bound size of the vertex buffer starting from the corresponding elements of pBuffers[i] plus pOffsets[i]. All vertex input attributes that use each of these bindings will use these updated addresses in their address calculations for subsequent draw commands.

If the bound pipeline state object was created with the DYNAMIC_STATE_VERTEX_INPUT_BINDING_STRIDE_EXT dynamic state enabled then pStrides[i] specifies the distance in bytes between two consecutive elements within the corresponding buffer. In this case the VertexInputBindingDescription::stride state from the pipeline state object is ignored.

Valid Usage

firstBinding must be less than PhysicalDeviceLimits::maxVertexInputBindings

The sum of firstBinding and bindingCount must be less than or equal to PhysicalDeviceLimits::maxVertexInputBindings

All elements of pOffsets must be less than the size of the corresponding element in pBuffers

If pSizes is not NULL, all elements of pOffsets plus pSizes must be less than or equal to the size of the corresponding element in pBuffers

All elements of pBuffers must have been created with the BUFFER_USAGE_VERTEX_BUFFER_BIT flag

Each element of pBuffers that is non-sparse must be bound completely and contiguously to a single DeviceMemory object

If the nullDescriptor feature is not enabled, all elements of pBuffers must not be NULL_HANDLE

If an element of pBuffers is NULL_HANDLE, then the corresponding element of pOffsets must be zero

If the bound pipeline state object was created with the DYNAMIC_STATE_VERTEX_INPUT_BINDING_STRIDE_EXT dynamic state enabled then pStrides must not be NULL, otherwise pStrides must be NULL

If pStrides is not NULL each element of pStrides must be less than or equal to PhysicalDeviceLimits::maxVertexInputBindingStride

If pStrides is not NULL each element of pStrides must be greater than or equal to the maximum extent of of all vertex input attributes fetched from the corresponding binding, where the extent is calculated as the VkVertexInputAttributeDescription::offset plus VkVertexInputAttributeDescription::format size

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pBuffers must be a valid pointer to an array of bindingCount valid Buffer handles

pOffsets must be a valid pointer to an array of bindingCount DeviceSize values

If pSizes is not NULL, pSizes must be a valid pointer to an array of bindingCount DeviceSize values

If pStrides is not NULL, pStrides must be a valid pointer to an array of bindingCount DeviceSize values

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

If any of pSizes, or pStrides are not NULL, bindingCount must be greater than 0

Both of commandBuffer, and the elements of pBuffers must have been created, allocated, or retrieved from the same Device

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

Buffer, CommandBuffer, DeviceSize

cmdSetDepthTestEnableEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command will be recorded.

-> ("depthTestEnable" ::: Bool)
depthTestEnable specifies if the depth test is enabled.

-> io ()

vkCmdSetDepthTestEnableEXT - Set the depth test enable for a command buffer

Description

This command sets the state for a given draw when the graphics pipeline is created with DYNAMIC_STATE_DEPTH_TEST_ENABLE_EXT set in PipelineDynamicStateCreateInfo::pDynamicStates.

Valid Usage

The extendedDynamicState feature must be enabled

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

Bool32, CommandBuffer

cmdSetDepthWriteEnableEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command will be recorded.

-> ("depthWriteEnable" ::: Bool)
depthWriteEnable specifies if depth writes are enabled.

-> io ()

vkCmdSetDepthWriteEnableEXT - Set the depth write enable for the command buffer

Description

This command sets the state for a given draw when the graphics pipeline is created with DYNAMIC_STATE_DEPTH_WRITE_ENABLE_EXT set in PipelineDynamicStateCreateInfo::pDynamicStates.

Valid Usage

The extendedDynamicState feature must be enabled

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

Bool32, CommandBuffer

cmdSetDepthCompareOpEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command will be recorded.

-> ("depthCompareOp" ::: CompareOp)
depthCompareOp specifies the depth comparison operator.

-> io ()

vkCmdSetDepthCompareOpEXT - Set the depth comparison operator for the command buffer

Description

This command sets the state for a given draw when the graphics pipeline is created with DYNAMIC_STATE_DEPTH_COMPARE_OP_EXT set in PipelineDynamicStateCreateInfo::pDynamicStates.

Valid Usage

The extendedDynamicState feature must be enabled

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

depthCompareOp must be a valid CompareOp value

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

CommandBuffer, CompareOp

cmdSetDepthBoundsTestEnableEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command will be recorded.

-> ("depthBoundsTestEnable" ::: Bool)
depthBoundsTestEnable specifies if the depth bounds test is enabled.

-> io ()

vkCmdSetDepthBoundsTestEnableEXT - Set the depth bounds test enable for a command buffer

Description

This command sets the state for a given draw when the graphics pipeline is created with DYNAMIC_STATE_DEPTH_BOUNDS_TEST_ENABLE_EXT set in PipelineDynamicStateCreateInfo::pDynamicStates.

Valid Usage

The extendedDynamicState feature must be enabled

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

Bool32, CommandBuffer

cmdSetStencilTestEnableEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command will be recorded.

-> ("stencilTestEnable" ::: Bool)
stencilTestEnable specifies if the stencil test is enabled.

-> io ()

vkCmdSetStencilTestEnableEXT - Set the stencil test enable for the command buffer

Description

This command sets the state for a given draw when the graphics pipeline is created with DYNAMIC_STATE_STENCIL_TEST_ENABLE_EXT set in PipelineDynamicStateCreateInfo::pDynamicStates.

Valid Usage

The extendedDynamicState feature must be enabled

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

Bool32, CommandBuffer

cmdSetStencilOpEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command will be recorded.

-> ("faceMask" ::: StencilFaceFlags)
faceMask is a bitmask of VkStencilFaceFlagBits specifying the set of stencil state for which to update the stencil operation.

-> ("failOp" ::: StencilOp)
failOp is a StencilOp value specifying the action performed on samples that fail the stencil test.

-> ("passOp" ::: StencilOp)
passOp is a StencilOp value specifying the action performed on samples that pass both the depth and stencil tests.

-> ("depthFailOp" ::: StencilOp)
depthFailOp is a StencilOp value specifying the action performed on samples that pass the stencil test and fail the depth test.

-> CompareOp
compareOp is a CompareOp value specifying the comparison operator used in the stencil test.

-> io ()

vkCmdSetStencilOpEXT - Set the stencil operation for the command buffer

Description

This command sets the state for a given draw when the graphics pipeline is created with DYNAMIC_STATE_STENCIL_OP_EXT set in PipelineDynamicStateCreateInfo::pDynamicStates.

Valid Usage

The extendedDynamicState feature must be enabled

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

faceMask must be a valid combination of StencilFaceFlagBits values

faceMask must not be 0

failOp must be a valid StencilOp value

passOp must be a valid StencilOp value

depthFailOp must be a valid StencilOp value

compareOp must be a valid CompareOp value

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

CommandBuffer, CompareOp, StencilFaceFlags, StencilOp

data PhysicalDeviceExtendedDynamicStateFeaturesEXT

VkPhysicalDeviceExtendedDynamicStateFeaturesEXT - Structure describing what extended dynamic state can be used

Members

The members of the PhysicalDeviceExtendedDynamicStateFeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceExtendedDynamicStateFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether the feature is supported. PhysicalDeviceExtendedDynamicStateFeaturesEXT can also be used in the pNext chain of DeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceExtendedDynamicStateFeaturesEXT

Fields

extendedDynamicState :: Bool

extendedDynamicState indicates that the implementation supports the following dynamic states:

DYNAMIC_STATE_CULL_MODE_EXT

DYNAMIC_STATE_FRONT_FACE_EXT

DYNAMIC_STATE_PRIMITIVE_TOPOLOGY_EXT

DYNAMIC_STATE_VIEWPORT_WITH_COUNT_EXT

DYNAMIC_STATE_SCISSOR_WITH_COUNT_EXT

DYNAMIC_STATE_VERTEX_INPUT_BINDING_STRIDE_EXT

DYNAMIC_STATE_DEPTH_TEST_ENABLE_EXT

DYNAMIC_STATE_DEPTH_WRITE_ENABLE_EXT

DYNAMIC_STATE_DEPTH_COMPARE_OP_EXT

DYNAMIC_STATE_DEPTH_BOUNDS_TEST_ENABLE_EXT

DYNAMIC_STATE_STENCIL_TEST_ENABLE_EXT

DYNAMIC_STATE_STENCIL_OP_EXT

type EXT_EXTENDED_DYNAMIC_STATE_SPEC_VERSION = 1

pattern EXT_EXTENDED_DYNAMIC_STATE_SPEC_VERSION :: forall a. Integral a => a

type EXT_EXTENDED_DYNAMIC_STATE_EXTENSION_NAME = "VK_EXT_extended_dynamic_state"

pattern EXT_EXTENDED_DYNAMIC_STATE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a