Vulkan.Extensions.VK_EXT_filter_cubic

Documentation

pattern FILTER_CUBIC_EXT :: Filter

pattern FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_CUBIC_BIT_EXT :: FormatFeatureFlagBits

data PhysicalDeviceImageViewImageFormatInfoEXT

VkPhysicalDeviceImageViewImageFormatInfoEXT - Structure for providing image view type

Valid Usage (Implicit)

See Also

ImageViewType, StructureType

Constructors

PhysicalDeviceImageViewImageFormatInfoEXT

Fields

imageViewType :: ImageViewType

imageViewType is a ImageViewType value specifying the type of the image view.

imageViewType must be a valid ImageViewType value

data FilterCubicImageViewImageFormatPropertiesEXT

VkFilterCubicImageViewImageFormatPropertiesEXT - Structure for querying cubic filtering capabilities of an image view type

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_FILTER_CUBIC_IMAGE_VIEW_IMAGE_FORMAT_PROPERTIES_EXT

Valid Usage

If the pNext chain of the ImageFormatProperties2 structure includes a FilterCubicImageViewImageFormatPropertiesEXT structure, the pNext chain of the PhysicalDeviceImageFormatInfo2 structure must include a PhysicalDeviceImageViewImageFormatInfoEXT structure with an imageViewType that is compatible with imageType

See Also

Bool32, StructureType

Constructors

FilterCubicImageViewImageFormatPropertiesEXT

Fields

filterCubic :: Bool

filterCubic tells if image format, image type and image view type can be used with cubic filtering. This field is set by the implementation. User-specified value is ignored.

filterCubicMinmax :: Bool

filterCubicMinmax tells if image format, image type and image view type can be used with cubic filtering and minmax filtering. This field is set by the implementation. User-specified value is ignored.

type EXT_FILTER_CUBIC_SPEC_VERSION = 3

pattern EXT_FILTER_CUBIC_SPEC_VERSION :: forall a. Integral a => a

type EXT_FILTER_CUBIC_EXTENSION_NAME = "VK_EXT_filter_cubic"

pattern EXT_FILTER_CUBIC_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a