Vulkan.Extensions.VK_EXT_fragment_density_map

Documentation

data PhysicalDeviceFragmentDensityMapFeaturesEXT

VkPhysicalDeviceFragmentDensityMapFeaturesEXT - Structure describing fragment density map features that can be supported by an implementation

Members

The members of the PhysicalDeviceFragmentDensityMapFeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceFragmentDensityMapFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether each feature is supported. PhysicalDeviceFragmentDensityMapFeaturesEXT can also be included in pNext chain of DeviceCreateInfo to enable the features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceFragmentDensityMapFeaturesEXT

Fields

fragmentDensityMap :: Bool

fragmentDensityMap specifies whether the implementation supports render passes with a fragment density map attachment. If this feature is not enabled and the pNext chain of RenderPassCreateInfo includes a RenderPassFragmentDensityMapCreateInfoEXT structure, fragmentDensityMapAttachment must be ATTACHMENT_UNUSED.

fragmentDensityMapDynamic :: Bool

fragmentDensityMapDynamic specifies whether the implementation supports dynamic fragment density map image views. If this feature is not enabled, IMAGE_VIEW_CREATE_FRAGMENT_DENSITY_MAP_DYNAMIC_BIT_EXT must not be included in ImageViewCreateInfo::flags.

fragmentDensityMapNonSubsampledImages :: Bool

fragmentDensityMapNonSubsampledImages specifies whether the implementation supports regular non-subsampled image attachments with fragment density map render passes. If this feature is not enabled, render passes with a fragment density map attachment must only have subsampled attachments bound.

data PhysicalDeviceFragmentDensityMapPropertiesEXT

VkPhysicalDeviceFragmentDensityMapPropertiesEXT - Structure describing fragment density map properties that can be supported by an implementation

Members

The members of the PhysicalDeviceFragmentDensityMapPropertiesEXT structure describe the following implementation-dependent limits:

Valid Usage (Implicit)

If the PhysicalDeviceFragmentDensityMapPropertiesEXT structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits and properties.

See Also

Bool32, Extent2D, StructureType

Constructors

PhysicalDeviceFragmentDensityMapPropertiesEXT

Fields

minFragmentDensityTexelSize :: Extent2D

minFragmentDensityTexelSize is the minimum fragment density texel size.

maxFragmentDensityTexelSize :: Extent2D

maxFragmentDensityTexelSize is the maximum fragment density texel size.

fragmentDensityInvocations :: Bool

fragmentDensityInvocations specifies whether the implementation may invoke additional fragment shader invocations for each covered sample.

data RenderPassFragmentDensityMapCreateInfoEXT

VkRenderPassFragmentDensityMapCreateInfoEXT - Structure containing fragment density map attachment for render pass

Description

The fragment density map attachment is read at an implementation-dependent time either by the host during cmdBeginRenderPass if the attachment’s image view was not created with flags containing IMAGE_VIEW_CREATE_FRAGMENT_DENSITY_MAP_DYNAMIC_BIT_EXT, or by the device when drawing commands in the renderpass execute PIPELINE_STAGE_FRAGMENT_DENSITY_PROCESS_BIT_EXT.

If this structure is not present, it is as if fragmentDensityMapAttachment was given as ATTACHMENT_UNUSED.

Valid Usage

If fragmentDensityMapAttachment is not ATTACHMENT_UNUSED, fragmentDensityMapAttachment must be less than RenderPassCreateInfo::attachmentCount

If fragmentDensityMapAttachment is not ATTACHMENT_UNUSED, fragmentDensityMapAttachment must not be an element of SubpassDescription::pInputAttachments, SubpassDescription::pColorAttachments, SubpassDescription::pResolveAttachments, SubpassDescription::pDepthStencilAttachment, or SubpassDescription::pPreserveAttachments for any subpass

If fragmentDensityMapAttachment is not ATTACHMENT_UNUSED, layout must be equal to IMAGE_LAYOUT_FRAGMENT_DENSITY_MAP_OPTIMAL_EXT, or IMAGE_LAYOUT_GENERAL

If fragmentDensityMapAttachment is not ATTACHMENT_UNUSED, fragmentDensityMapAttachment must reference an attachment with a loadOp equal to ATTACHMENT_LOAD_OP_LOAD or ATTACHMENT_LOAD_OP_DONT_CARE

If fragmentDensityMapAttachment is not ATTACHMENT_UNUSED, fragmentDensityMapAttachment must reference an attachment with a storeOp equal to ATTACHMENT_STORE_OP_DONT_CARE

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_RENDER_PASS_FRAGMENT_DENSITY_MAP_CREATE_INFO_EXT

fragmentDensityMapAttachment must be a valid AttachmentReference structure

See Also

AttachmentReference, StructureType

Constructors

RenderPassFragmentDensityMapCreateInfoEXT

Fields

fragmentDensityMapAttachment :: AttachmentReference

fragmentDensityMapAttachment is the fragment density map to use for the render pass.

type EXT_FRAGMENT_DENSITY_MAP_SPEC_VERSION = 1

pattern EXT_FRAGMENT_DENSITY_MAP_SPEC_VERSION :: forall a. Integral a => a

type EXT_FRAGMENT_DENSITY_MAP_EXTENSION_NAME = "VK_EXT_fragment_density_map"

pattern EXT_FRAGMENT_DENSITY_MAP_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a