Vulkan.Extensions.VK_EXT_fragment_density_map2

Documentation

data PhysicalDeviceFragmentDensityMap2FeaturesEXT

VkPhysicalDeviceFragmentDensityMap2FeaturesEXT - Structure describing additional fragment density map features that can be supported by an implementation

Members

The members of the PhysicalDeviceFragmentDensityMap2FeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceFragmentDensityMap2FeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether each feature is supported. PhysicalDeviceFragmentDensityMap2FeaturesEXT can also be included in pNext chain of DeviceCreateInfo to enable the features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceFragmentDensityMap2FeaturesEXT

Fields

fragmentDensityMapDeferred :: Bool

fragmentDensityMapDeferred specifies whether the implementation supports deferred reads of fragment density map image views. If this feature is not enabled, IMAGE_VIEW_CREATE_FRAGMENT_DENSITY_MAP_DEFERRED_BIT_EXT must not be included in ImageViewCreateInfo::flags.

data PhysicalDeviceFragmentDensityMap2PropertiesEXT

VkPhysicalDeviceFragmentDensityMap2PropertiesEXT - Structure describing additional fragment density map properties that can be supported by an implementation

Members

The members of the PhysicalDeviceFragmentDensityMap2PropertiesEXT structure describe the following implementation-dependent limits:

Valid Usage (Implicit)

If the PhysicalDeviceFragmentDensityMap2PropertiesEXT structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits and properties.

See Also

Bool32, StructureType

Constructors

PhysicalDeviceFragmentDensityMap2PropertiesEXT

Fields

subsampledLoads :: Bool

subsampledLoads specifies if performing image data read with load operations on subsampled attachments will be resampled to the fragment density of the render pass

subsampledCoarseReconstructionEarlyAccess :: Bool

subsampledCoarseReconstructionEarlyAccess specifies if performing image data read with samplers created with flags containing SAMPLER_CREATE_SUBSAMPLED_COARSE_RECONSTRUCTION_BIT_EXT in fragment shader will trigger additional reads during PIPELINE_STAGE_VERTEX_SHADER_BIT

maxSubsampledArrayLayers :: Word32

maxSubsampledArrayLayers is the maximum number of ImageView array layers for usages supporting subsampled samplers

maxDescriptorSetSubsampledSamplers :: Word32

maxDescriptorSetSubsampledSamplers is the maximum number of subsampled samplers that can be included in a PipelineLayout

type EXT_FRAGMENT_DENSITY_MAP_2_SPEC_VERSION = 1

pattern EXT_FRAGMENT_DENSITY_MAP_2_SPEC_VERSION :: forall a. Integral a => a

type EXT_FRAGMENT_DENSITY_MAP_2_EXTENSION_NAME = "VK_EXT_fragment_density_map2"

pattern EXT_FRAGMENT_DENSITY_MAP_2_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a