Vulkan.Extensions.VK_EXT_host_query_reset

Description

Name

VK_EXT_host_query_reset - device extension

VK_EXT_host_query_reset

Name String

VK_EXT_host_query_reset

Extension Type

Device extension

Registered Extension Number

262

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_get_physical_device_properties2

Deprecation state

Promoted to Vulkan 1.2

Contact

Bas Nieuwenhuizen https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_EXT_host_query_reset:%20&body=@BNieuwenhuizen%20

Other Extension Metadata

Last Modified Date

2019-03-06

IP Status

No known IP claims.

Interactions and External Dependencies

Promoted to Vulkan 1.2 Core

Contributors

Bas Nieuwenhuizen, Google

Jason Ekstrand, Intel

Jeff Bolz, NVIDIA

Piers Daniell, NVIDIA

Description

This extension adds a new function to reset queries from the host.

Promotion to Vulkan 1.2

All functionality in this extension is included in core Vulkan 1.2, with the EXT suffix omitted. The original type, enum and command names are still available as aliases of the core functionality.

New Commands

resetQueryPoolEXT

New Structures

Extending PhysicalDeviceFeatures2, DeviceCreateInfo:

PhysicalDeviceHostQueryResetFeaturesEXT

New Enum Constants

EXT_HOST_QUERY_RESET_EXTENSION_NAME

EXT_HOST_QUERY_RESET_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_PHYSICAL_DEVICE_HOST_QUERY_RESET_FEATURES_EXT

Version History

Revision 1, 2019-03-12 (Bas Nieuwenhuizen)

Initial draft

See Also

PhysicalDeviceHostQueryResetFeaturesEXT, resetQueryPoolEXT

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_HOST_QUERY_RESET_FEATURES_EXT :: StructureType

resetQueryPoolEXT :: MonadIO io => Device -> QueryPool -> ("firstQuery" ::: Word32) -> ("queryCount" ::: Word32) -> io ()

type PhysicalDeviceHostQueryResetFeaturesEXT = PhysicalDeviceHostQueryResetFeatures

type EXT_HOST_QUERY_RESET_SPEC_VERSION = 1

pattern EXT_HOST_QUERY_RESET_SPEC_VERSION :: forall a. Integral a => a

type EXT_HOST_QUERY_RESET_EXTENSION_NAME = "VK_EXT_host_query_reset"

pattern EXT_HOST_QUERY_RESET_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a