Vulkan.Extensions.VK_EXT_inline_uniform_block

Description

Name

VK_EXT_inline_uniform_block - device extension

VK_EXT_inline_uniform_block

Name String

VK_EXT_inline_uniform_block

Extension Type

Device extension

Registered Extension Number

139

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_get_physical_device_properties2

Requires VK_KHR_maintenance1

Contact

Daniel Rakos https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_EXT_inline_uniform_block:%20&body=@aqnuep%20

Other Extension Metadata

Last Modified Date

2018-08-01

IP Status

No known IP claims.

Contributors

Daniel Rakos, AMD

Jeff Bolz, NVIDIA

Slawomir Grajewski, Intel

Neil Henning, Codeplay

Description

This extension introduces the ability to back uniform blocks directly with descriptor sets by storing inline uniform data within descriptor pool storage. Compared to push constants this new construct allows uniform data to be reused across multiple disjoint sets of draw or dispatch commands and may enable uniform data to be accessed with less indirections compared to uniforms backed by buffer memory.

New Structures

Extending DescriptorPoolCreateInfo:

DescriptorPoolInlineUniformBlockCreateInfoEXT

Extending PhysicalDeviceFeatures2, DeviceCreateInfo:

PhysicalDeviceInlineUniformBlockFeaturesEXT

Extending PhysicalDeviceProperties2:

PhysicalDeviceInlineUniformBlockPropertiesEXT

Extending WriteDescriptorSet:

WriteDescriptorSetInlineUniformBlockEXT

New Enum Constants

EXT_INLINE_UNIFORM_BLOCK_EXTENSION_NAME

EXT_INLINE_UNIFORM_BLOCK_SPEC_VERSION

Extending DescriptorType:

DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT

Extending StructureType:

STRUCTURE_TYPE_DESCRIPTOR_POOL_INLINE_UNIFORM_BLOCK_CREATE_INFO_EXT

STRUCTURE_TYPE_PHYSICAL_DEVICE_INLINE_UNIFORM_BLOCK_FEATURES_EXT

STRUCTURE_TYPE_PHYSICAL_DEVICE_INLINE_UNIFORM_BLOCK_PROPERTIES_EXT

STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET_INLINE_UNIFORM_BLOCK_EXT

Issues

1) Do we need a new storage class for inline uniform blocks vs uniform blocks?

RESOLVED: No. The Uniform storage class is used to allow the same syntax used for both uniform buffers and inline uniform blocks.

2) Is the descriptor array index and array size expressed in terms of bytes or dwords for inline uniform block descriptors?

RESOLVED: In bytes, but both must be a multiple of 4, similar to how push constant ranges are specified. The descriptorCount of DescriptorSetLayoutBinding thus provides the total number of bytes a particular binding with an inline uniform block descriptor type can hold, while the srcArrayElement, dstArrayElement, and descriptorCount members of WriteDescriptorSet, CopyDescriptorSet, and DescriptorUpdateTemplateEntry (where applicable) specify the byte offset and number of bytes to write/copy to the binding’s backing store. Additionally, the stride member of DescriptorUpdateTemplateEntry is ignored for inline uniform blocks and a default value of one is used, meaning that the data to update inline uniform block bindings with must be contiguous in memory.

3) What layout rules apply for uniform blocks corresponding to inline constants?

RESOLVED: They use the same layout rules as uniform buffers.

4) Do we need to add non-uniform indexing features/properties as introduced by VK_EXT_descriptor_indexing for inline uniform blocks?

RESOLVED: No, because inline uniform blocks are not allowed to be “arrayed”. A single binding with an inline uniform block descriptor type corresponds to a single uniform block instance and the array indices inside that binding refer to individual offsets within the uniform block (see issue #2). However, this extension does introduce new features/properties about the level of support for update-after-bind inline uniform blocks.

5) Is the descriptorBindingVariableDescriptorCount feature introduced by VK_EXT_descriptor_indexing supported for inline uniform blocks?

RESOLVED: Yes, as long as other inline uniform block specific limits are respected.

6) Do the robustness guarantees of robustBufferAccess apply to inline uniform block accesses?

RESOLVED: No, similarly to push constants, as they are not backed by buffer memory like uniform buffers.

Version History

Revision 1, 2018-08-01 (Daniel Rakos)

Internal revisions

See Also

DescriptorPoolInlineUniformBlockCreateInfoEXT, PhysicalDeviceInlineUniformBlockFeaturesEXT, PhysicalDeviceInlineUniformBlockPropertiesEXT, WriteDescriptorSetInlineUniformBlockEXT

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data PhysicalDeviceInlineUniformBlockFeaturesEXT

VkPhysicalDeviceInlineUniformBlockFeaturesEXT - Structure describing inline uniform block features that can be supported by an implementation

Members

The members of the PhysicalDeviceInlineUniformBlockFeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceInlineUniformBlockFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether each feature is supported. PhysicalDeviceInlineUniformBlockFeaturesEXT can also be included in the pNext chain of DeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceInlineUniformBlockFeaturesEXT

Fields

inlineUniformBlock :: Bool

inlineUniformBlock indicates whether the implementation supports inline uniform block descriptors. If this feature is not enabled, DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT must not be used.

descriptorBindingInlineUniformBlockUpdateAfterBind :: Bool

descriptorBindingInlineUniformBlockUpdateAfterBind indicates whether the implementation supports updating inline uniform block descriptors after a set is bound. If this feature is not enabled, DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT must not be used with DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT.

data PhysicalDeviceInlineUniformBlockPropertiesEXT

VkPhysicalDeviceInlineUniformBlockPropertiesEXT - Structure describing inline uniform block properties that can be supported by an implementation

Members

The members of the PhysicalDeviceInlineUniformBlockPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the PhysicalDeviceInlineUniformBlockPropertiesEXT structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

StructureType

Constructors

PhysicalDeviceInlineUniformBlockPropertiesEXT

Fields

maxInlineUniformBlockSize :: Word32

maxInlineUniformBlockSize is the maximum size in bytes of an inline uniform block binding.

maxPerStageDescriptorInlineUniformBlocks :: Word32

maxPerStageDescriptorUpdateAfterBindInlineUniformBlocks :: Word32

maxPerStageDescriptorUpdateAfterBindInlineUniformBlocks is similar to maxPerStageDescriptorInlineUniformBlocks but counts descriptor bindings from descriptor sets created with or without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set.

maxDescriptorSetInlineUniformBlocks :: Word32

maxDescriptorSetInlineUniformBlocks is the maximum number of inline uniform block bindings that can be included in descriptor bindings in a pipeline layout across all pipeline shader stages and descriptor set numbers. Descriptor bindings with a descriptor type of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT count against this limit. Only descriptor bindings in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set count against this limit.

maxDescriptorSetUpdateAfterBindInlineUniformBlocks :: Word32

maxDescriptorSetUpdateAfterBindInlineUniformBlocks is similar to maxDescriptorSetInlineUniformBlocks but counts descriptor bindings from descriptor sets created with or without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set.

data WriteDescriptorSetInlineUniformBlockEXT

VkWriteDescriptorSetInlineUniformBlockEXT - Structure specifying inline uniform block data

Valid Usage (Implicit)

See Also

StructureType

Constructors

WriteDescriptorSetInlineUniformBlockEXT

Fields

dataSize :: Word32

dataSize is the number of bytes of inline uniform block data pointed to by pData.

dataSize must be an integer multiple of 4

dataSize must be greater than 0

data' :: Ptr ()

pData is a pointer to dataSize number of bytes of data to write to the inline uniform block.

pData must be a valid pointer to an array of dataSize bytes

data DescriptorPoolInlineUniformBlockCreateInfoEXT

VkDescriptorPoolInlineUniformBlockCreateInfoEXT - Structure specifying the maximum number of inline uniform block bindings of a newly created descriptor pool

Valid Usage (Implicit)

See Also

StructureType

Constructors

DescriptorPoolInlineUniformBlockCreateInfoEXT

Fields

maxInlineUniformBlockBindings :: Word32

maxInlineUniformBlockBindings is the number of inline uniform block bindings to allocate.

type EXT_INLINE_UNIFORM_BLOCK_SPEC_VERSION = 1

pattern EXT_INLINE_UNIFORM_BLOCK_SPEC_VERSION :: forall a. Integral a => a

type EXT_INLINE_UNIFORM_BLOCK_EXTENSION_NAME = "VK_EXT_inline_uniform_block"

pattern EXT_INLINE_UNIFORM_BLOCK_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a