Vulkan.Extensions.VK_EXT_memory_priority

Documentation

data PhysicalDeviceMemoryPriorityFeaturesEXT

VkPhysicalDeviceMemoryPriorityFeaturesEXT - Structure describing memory priority features that can be supported by an implementation

Members

The members of the PhysicalDeviceMemoryPriorityFeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceMemoryPriorityFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether the feature is supported. PhysicalDeviceMemoryPriorityFeaturesEXT can also be included in the pNext chain of DeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceMemoryPriorityFeaturesEXT

Fields

memoryPriority :: Bool

memoryPriority indicates that the implementation supports memory priorities specified at memory allocation time via MemoryPriorityAllocateInfoEXT.

data MemoryPriorityAllocateInfoEXT

VkMemoryPriorityAllocateInfoEXT - Specify a memory allocation priority

Description

Memory allocations with higher priority may be more likely to stay in device-local memory when the system is under memory pressure.

If this structure is not included, it is as if the priority value were 0.5.

Valid Usage (Implicit)

See Also

StructureType

Constructors

MemoryPriorityAllocateInfoEXT

Fields

priority :: Float

priority is a floating-point value between 0 and 1, indicating the priority of the allocation relative to other memory allocations. Larger values are higher priority. The granularity of the priorities is implementation-dependent.

priority must be between 0 and 1, inclusive

type EXT_MEMORY_PRIORITY_SPEC_VERSION = 1

pattern EXT_MEMORY_PRIORITY_SPEC_VERSION :: forall a. Integral a => a

type EXT_MEMORY_PRIORITY_EXTENSION_NAME = "VK_EXT_memory_priority"

pattern EXT_MEMORY_PRIORITY_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a