Vulkan.Extensions.VK_EXT_private_data

Documentation

createPrivateDataSlotEXT

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device associated with the creation of the object(s) holding the private data slot.

-> PrivateDataSlotCreateInfoEXT
pCreateInfo is a pointer to a PrivateDataSlotCreateInfoEXT

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io PrivateDataSlotEXT

vkCreatePrivateDataSlotEXT - Create a slot for private data storage

Valid Usage (Implicit)

device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid PrivateDataSlotCreateInfoEXT structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pPrivateDataSlot must be a valid pointer to a PrivateDataSlotEXT handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

See Also

AllocationCallbacks, Device, PrivateDataSlotCreateInfoEXT, PrivateDataSlotEXT

withPrivateDataSlotEXT :: forall io r. MonadIO io => Device -> PrivateDataSlotCreateInfoEXT -> Maybe AllocationCallbacks -> (io PrivateDataSlotEXT -> (PrivateDataSlotEXT -> io ()) -> r) -> r

A convenience wrapper to make a compatible pair of calls to createPrivateDataSlotEXT and destroyPrivateDataSlotEXT

To ensure that destroyPrivateDataSlotEXT is always called: pass bracket (or the allocate function from your favourite resource management library) as the first argument. To just extract the pair pass (,) as the first argument.

destroyPrivateDataSlotEXT

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device associated with the creation of the object(s) holding the private data slot.

-> PrivateDataSlotEXT
privateDataSlot is the private data slot to destroy.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io ()

vkDestroyPrivateDataSlotEXT - Destroy a private data slot

See Also

AllocationCallbacks, Device, PrivateDataSlotEXT

setPrivateDataEXT

Arguments

:: forall io. MonadIO io

=> Device
device is the device that created the object.

device must be a valid Device handle

-> ObjectType
objectType is a ObjectType specifying the type of object to associate data with.

objectType must be a valid ObjectType value

-> ("objectHandle" ::: Word64)
objectHandle is a handle to the object to associate data with.

objectHandle must be device or a child of device

objectHandle must be a valid handle to an object of type objectType

-> PrivateDataSlotEXT
privateDataSlot is a handle to a PrivateDataSlotEXT specifying location of private data storage.

privateDataSlot must be a valid PrivateDataSlotEXT handle

privateDataSlot must have been created, allocated, or retrieved from device

-> ("data" ::: Word64)
data is user defined data to associate the object with. This data will be stored at privateDataSlot.

-> io ()

vkSetPrivateDataEXT - Associate data with a Vulkan object

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

See Also

Device, ObjectType, PrivateDataSlotEXT

getPrivateDataEXT

Arguments

:: forall io. MonadIO io

=> Device
device is the device that created the object

device must be a valid Device handle

-> ObjectType
objectType is a ObjectType specifying the type of object data is associated with.

objectType must be Device or an object type whose parent is Device

objectType must be a valid ObjectType value

-> ("objectHandle" ::: Word64)
objectHandle is a handle to the object data is associated with.

-> PrivateDataSlotEXT
privateDataSlot is a handle to a PrivateDataSlotEXT specifying location of private data pointer storage.

privateDataSlot must be a valid PrivateDataSlotEXT handle

privateDataSlot must have been created, allocated, or retrieved from device

-> io ("data" ::: Word64)

vkGetPrivateDataEXT - Retrieve data associated with a Vulkan object

Description

Note

Due to platform details on Android, implementations might not be able to reliably return 0 from calls to getPrivateDataEXT for SwapchainKHR objects on which setPrivateDataEXT has not previously been called. This erratum is exclusive to the Android platform and objects of type SwapchainKHR.

Valid Usage (Implicit)

See Also

Device, ObjectType, PrivateDataSlotEXT

data DevicePrivateDataCreateInfoEXT

VkDevicePrivateDataCreateInfoEXT - Reserve private data slots

Valid Usage (Implicit)

See Also

StructureType

Constructors

DevicePrivateDataCreateInfoEXT

Fields

privateDataSlotRequestCount :: Word32

privateDataSlotRequestCount is the amount of slots to reserve.

data PrivateDataSlotCreateInfoEXT

VkPrivateDataSlotCreateInfoEXT - Structure specifying the parameters of private data slot construction

Valid Usage (Implicit)

See Also

PrivateDataSlotCreateFlagsEXT, StructureType, createPrivateDataSlotEXT

Constructors

PrivateDataSlotCreateInfoEXT

Fields

flags :: PrivateDataSlotCreateFlagsEXT

flags is a bitmask of PrivateDataSlotCreateFlagsEXT specifying additional parameters of the new private data slot

flags must be 0

data PhysicalDevicePrivateDataFeaturesEXT

VkPhysicalDevicePrivateDataFeaturesEXT - Structure specifying physical device support

Members

The members of the PhysicalDevicePrivateDataFeaturesEXT structure describe the following features:

Description

If the PhysicalDevicePrivateDataFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether the feature is supported. PhysicalDevicePrivateDataFeaturesEXT can also be used in the pNext chain of DeviceCreateInfo to enable the features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDevicePrivateDataFeaturesEXT

Fields

privateData :: Bool

privateData indicates whether the implementation supports private data. See Private Data.

newtype PrivateDataSlotCreateFlagBitsEXT

VkPrivateDataSlotCreateFlagBitsEXT - Bitmask specifying additional parameters for private data slot creation

See Also

PrivateDataSlotCreateFlagsEXT

Constructors

PrivateDataSlotCreateFlagBitsEXT Flags

type PrivateDataSlotCreateFlagsEXT = PrivateDataSlotCreateFlagBitsEXT

type EXT_PRIVATE_DATA_SPEC_VERSION = 1

pattern EXT_PRIVATE_DATA_SPEC_VERSION :: forall a. Integral a => a

type EXT_PRIVATE_DATA_EXTENSION_NAME = "VK_EXT_private_data"

pattern EXT_PRIVATE_DATA_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype PrivateDataSlotEXT

VkPrivateDataSlotEXT - Opaque handle to a private data slot object

See Also

createPrivateDataSlotEXT, destroyPrivateDataSlotEXT, getPrivateDataEXT, setPrivateDataEXT

Constructors

PrivateDataSlotEXT Word64