Vulkan.Extensions.VK_EXT_robustness2

Documentation

data PhysicalDeviceRobustness2FeaturesEXT

VkPhysicalDeviceRobustness2FeaturesEXT - Structure describing the out-of-bounds behavior for an implementation

Members

The members of the PhysicalDeviceRobustness2FeaturesEXT structure describe the following features:

Description

robustBufferAccess2 indicates whether buffer accesses are tightly bounds-checked against the range of the descriptor. Uniform buffers must be bounds-checked to the range of the descriptor, where the range is rounded up to a multiple of robustUniformBufferAccessSizeAlignment. Storage buffers must be bounds-checked to the range of the descriptor, where the range is rounded up to a multiple of robustStorageBufferAccessSizeAlignment. Out of bounds buffer loads will return zero values, and formatted loads will have (0,0,1) values inserted for missing G, B, or A components based on the format.

robustImageAccess2 indicates whether image accesses are tightly bounds-checked against the dimensions of the image view. Out of bounds image loads will return zero values, with (0,0,1) values inserted for missing G, B, or A components based on the format.

nullDescriptor indicates whether descriptors can be written with a NULL_HANDLE resource or view, which are considered valid to access and act as if the descriptor were bound to nothing.

If the PhysicalDeviceRobustness2FeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether each feature is supported.

Valid Usage

If robustBufferAccess2 is enabled then robustBufferAccess must also be enabled

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PHYSICAL_DEVICE_ROBUSTNESS_2_FEATURES_EXT

See Also

Bool32, StructureType

Constructors

PhysicalDeviceRobustness2FeaturesEXT

Fields

robustBufferAccess2 :: Bool

robustImageAccess2 :: Bool

nullDescriptor :: Bool

data PhysicalDeviceRobustness2PropertiesEXT

VkPhysicalDeviceRobustness2PropertiesEXT - Structure describing robust buffer access properties supported by an implementation

Members

The members of the PhysicalDeviceRobustness2PropertiesEXT structure describe the following implementation-dependent limits:

Description

If the PhysicalDeviceRobustness2PropertiesEXT structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

DeviceSize, StructureType

Constructors

PhysicalDeviceRobustness2PropertiesEXT

Fields

robustStorageBufferAccessSizeAlignment :: DeviceSize

robustStorageBufferAccessSizeAlignment is the number of bytes that the range of a storage buffer descriptor is rounded up to when used for bounds-checking when robustBufferAccess2 is enabled. This value is either 1 or 4.

robustUniformBufferAccessSizeAlignment :: DeviceSize

robustUniformBufferAccessSizeAlignment is the number of bytes that the range of a uniform buffer descriptor is rounded up to when used for bounds-checking when robustBufferAccess2 is enabled. This value is a power of two in the range [1, 256].

type EXT_ROBUSTNESS_2_SPEC_VERSION = 1

pattern EXT_ROBUSTNESS_2_SPEC_VERSION :: forall a. Integral a => a

type EXT_ROBUSTNESS_2_EXTENSION_NAME = "VK_EXT_robustness2"

pattern EXT_ROBUSTNESS_2_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a