Vulkan.Extensions.VK_EXT_sampler_filter_minmax

Description

Name

VK_EXT_sampler_filter_minmax - device extension

VK_EXT_sampler_filter_minmax

Name String

VK_EXT_sampler_filter_minmax

Extension Type

Device extension

Registered Extension Number

131

Revision

2

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_get_physical_device_properties2

Deprecation state

Promoted to Vulkan 1.2

Contact

Jeff Bolz https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_EXT_sampler_filter_minmax:%20&body=@jeffbolznv%20

Other Extension Metadata

Last Modified Date

2017-05-19

Interactions and External Dependencies

Promoted to Vulkan 1.2 Core

IP Status

No known IP claims.

Contributors

Jeff Bolz, NVIDIA

Piers Daniell, NVIDIA

Description

In unextended Vulkan, minification and magnification filters such as LINEAR allow sampled image lookups to return a filtered texel value produced by computing a weighted average of a collection of texels in the neighborhood of the texture coordinate provided.

This extension provides a new sampler parameter which allows applications to produce a filtered texel value by computing a component-wise minimum (MIN) or maximum (MAX) of the texels that would normally be averaged. The reduction mode is orthogonal to the minification and magnification filter parameters. The filter parameters are used to identify the set of texels used to produce a final filtered value; the reduction mode identifies how these texels are combined.

Promotion to Vulkan 1.2

All functionality in this extension is included in core Vulkan 1.2, with the EXT suffix omitted. The original type, enum and command names are still available as aliases of the core functionality.

New Structures

Extending PhysicalDeviceProperties2:

PhysicalDeviceSamplerFilterMinmaxPropertiesEXT

Extending SamplerCreateInfo:

SamplerReductionModeCreateInfoEXT

New Enums

SamplerReductionModeEXT

New Enum Constants

EXT_SAMPLER_FILTER_MINMAX_EXTENSION_NAME

EXT_SAMPLER_FILTER_MINMAX_SPEC_VERSION

Extending FormatFeatureFlagBits:

FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_MINMAX_BIT_EXT

Extending SamplerReductionMode:

SAMPLER_REDUCTION_MODE_MAX_EXT

SAMPLER_REDUCTION_MODE_MIN_EXT

SAMPLER_REDUCTION_MODE_WEIGHTED_AVERAGE_EXT

Extending StructureType:

STRUCTURE_TYPE_PHYSICAL_DEVICE_SAMPLER_FILTER_MINMAX_PROPERTIES_EXT

STRUCTURE_TYPE_SAMPLER_REDUCTION_MODE_CREATE_INFO_EXT

Version History

Revision 2, 2017-05-19 (Piers Daniell)

Renamed to EXT

Revision 1, 2017-03-25 (Jeff Bolz)

Internal revisions

See Also

PhysicalDeviceSamplerFilterMinmaxPropertiesEXT, SamplerReductionModeCreateInfoEXT, SamplerReductionModeEXT

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SAMPLER_FILTER_MINMAX_PROPERTIES_EXT :: StructureType

pattern STRUCTURE_TYPE_SAMPLER_REDUCTION_MODE_CREATE_INFO_EXT :: StructureType

pattern FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_MINMAX_BIT_EXT :: FormatFeatureFlagBits

pattern SAMPLER_REDUCTION_MODE_WEIGHTED_AVERAGE_EXT :: SamplerReductionMode

pattern SAMPLER_REDUCTION_MODE_MIN_EXT :: SamplerReductionMode

pattern SAMPLER_REDUCTION_MODE_MAX_EXT :: SamplerReductionMode

type SamplerReductionModeEXT = SamplerReductionMode

type PhysicalDeviceSamplerFilterMinmaxPropertiesEXT = PhysicalDeviceSamplerFilterMinmaxProperties

type SamplerReductionModeCreateInfoEXT = SamplerReductionModeCreateInfo

type EXT_SAMPLER_FILTER_MINMAX_SPEC_VERSION = 2

pattern EXT_SAMPLER_FILTER_MINMAX_SPEC_VERSION :: forall a. Integral a => a

type EXT_SAMPLER_FILTER_MINMAX_EXTENSION_NAME = "VK_EXT_sampler_filter_minmax"

pattern EXT_SAMPLER_FILTER_MINMAX_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a