Vulkan.Extensions.VK_EXT_texel_buffer_alignment

Documentation

data PhysicalDeviceTexelBufferAlignmentFeaturesEXT

VkPhysicalDeviceTexelBufferAlignmentFeaturesEXT - Structure describing the texel buffer alignment features that can be supported by an implementation

Members

The members of the PhysicalDeviceTexelBufferAlignmentFeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceTexelBufferAlignmentFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether the feature is supported. PhysicalDeviceTexelBufferAlignmentFeaturesEXT can also be included in the pNext chain of DeviceCreateInfo to enable the feature.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceTexelBufferAlignmentFeaturesEXT

Fields

texelBufferAlignment :: Bool

texelBufferAlignment indicates whether the implementation uses more specific alignment requirements advertised in PhysicalDeviceTexelBufferAlignmentPropertiesEXT rather than PhysicalDeviceLimits::minTexelBufferOffsetAlignment.

data PhysicalDeviceTexelBufferAlignmentPropertiesEXT

VkPhysicalDeviceTexelBufferAlignmentPropertiesEXT - Structure describing the texel buffer alignment requirements supported by an implementation

Members

The members of the PhysicalDeviceTexelBufferAlignmentPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the PhysicalDeviceTexelBufferAlignmentPropertiesEXT structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

If the single texel alignment property is FALSE, then the buffer view’s offset must be aligned to the corresponding byte alignment value. If the single texel alignment property is TRUE, then the buffer view’s offset must be aligned to the lesser of the corresponding byte alignment value or the size of a single texel, based on BufferViewCreateInfo::format. If the size of a single texel is a multiple of three bytes, then the size of a single component of the format is used instead.

These limits must not advertise a larger alignment than the required maximum minimum value of PhysicalDeviceLimits::minTexelBufferOffsetAlignment, for any format that supports use as a texel buffer.

Valid Usage (Implicit)

See Also

Bool32, DeviceSize, StructureType

Constructors

PhysicalDeviceTexelBufferAlignmentPropertiesEXT

Fields

storageTexelBufferOffsetAlignmentBytes :: DeviceSize

storageTexelBufferOffsetAlignmentBytes is a byte alignment that is sufficient for a storage texel buffer of any format.

storageTexelBufferOffsetSingleTexelAlignment :: Bool

storageTexelBufferOffsetSingleTexelAlignment indicates whether single texel alignment is sufficient for a storage texel buffer of any format.

uniformTexelBufferOffsetAlignmentBytes :: DeviceSize

uniformTexelBufferOffsetAlignmentBytes is a byte alignment that is sufficient for a uniform texel buffer of any format.

uniformTexelBufferOffsetSingleTexelAlignment :: Bool

uniformTexelBufferOffsetSingleTexelAlignment indicates whether single texel alignment is sufficient for a uniform texel buffer of any format.

type EXT_TEXEL_BUFFER_ALIGNMENT_SPEC_VERSION = 1

pattern EXT_TEXEL_BUFFER_ALIGNMENT_SPEC_VERSION :: forall a. Integral a => a

type EXT_TEXEL_BUFFER_ALIGNMENT_EXTENSION_NAME = "VK_EXT_texel_buffer_alignment"

pattern EXT_TEXEL_BUFFER_ALIGNMENT_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a