Vulkan.Extensions.VK_EXT_tooling_info

Documentation

getPhysicalDeviceToolPropertiesEXT

Arguments

:: forall io. MonadIO io

=> PhysicalDevice
physicalDevice is the handle to the physical device to query for active tools.

-> io (Result, "toolProperties" ::: Vector PhysicalDeviceToolPropertiesEXT)

vkGetPhysicalDeviceToolPropertiesEXT - Reports properties of tools active on the specified physical device

Description

If pToolProperties is NULL, then the number of tools currently active on physicalDevice is returned in pToolCount. Otherwise, pToolCount must point to a variable set by the user to the number of elements in the pToolProperties array, and on return the variable is overwritten with the number of structures actually written to pToolProperties. If pToolCount is less than the number of currently active tools, at most pToolCount structures will be written.

The count and properties of active tools may change in response to events outside the scope of the specification. An application should assume these properties might change at any given time.

Valid Usage (Implicit)

physicalDevice must be a valid PhysicalDevice handle

pToolCount must be a valid pointer to a uint32_t value

If the value referenced by pToolCount is not 0, and pToolProperties is not NULL, pToolProperties must be a valid pointer to an array of pToolCount PhysicalDeviceToolPropertiesEXT structures

Return Codes

Success

SUCCESS

INCOMPLETE

See Also

PhysicalDevice, PhysicalDeviceToolPropertiesEXT

data PhysicalDeviceToolPropertiesEXT

VkPhysicalDeviceToolPropertiesEXT - Structure providing information about an active tool

Valid Usage (Implicit)

See Also

StructureType, ToolPurposeFlagsEXT, getPhysicalDeviceToolPropertiesEXT

Constructors

PhysicalDeviceToolPropertiesEXT

Fields

name :: ByteString

name is a null-terminated UTF-8 string containing the name of the tool.

version :: ByteString

version is a null-terminated UTF-8 string containing the version of the tool.

purposes :: ToolPurposeFlagsEXT

purposes is a bitmask of ToolPurposeFlagBitsEXT which is populated with purposes supported by the tool.

description :: ByteString

description is a null-terminated UTF-8 string containing a description of the tool.

layer :: ByteString

layer is a null-terminated UTF-8 string that contains the name of the layer implementing the tool, if the tool is implemented in a layer - otherwise it may be an empty string.

newtype ToolPurposeFlagBitsEXT

VkToolPurposeFlagBitsEXT - Bitmask specifying the purposes of an active tool

See Also

ToolPurposeFlagsEXT

Constructors

ToolPurposeFlagBitsEXT Flags

Bundled Patterns

pattern TOOL_PURPOSE_VALIDATION_BIT_EXT :: ToolPurposeFlagBitsEXT
TOOL_PURPOSE_VALIDATION_BIT_EXT specifies that the tool provides validation of API usage.

pattern TOOL_PURPOSE_PROFILING_BIT_EXT :: ToolPurposeFlagBitsEXT
TOOL_PURPOSE_PROFILING_BIT_EXT specifies that the tool provides profiling of API usage.

pattern TOOL_PURPOSE_TRACING_BIT_EXT :: ToolPurposeFlagBitsEXT
TOOL_PURPOSE_TRACING_BIT_EXT specifies that the tool is capturing data about the application’s API usage, including anything from simple logging to capturing data for later replay.

pattern TOOL_PURPOSE_ADDITIONAL_FEATURES_BIT_EXT :: ToolPurposeFlagBitsEXT
TOOL_PURPOSE_ADDITIONAL_FEATURES_BIT_EXT specifies that the tool provides additional API features/extensions on top of the underlying implementation.

pattern TOOL_PURPOSE_MODIFYING_FEATURES_BIT_EXT :: ToolPurposeFlagBitsEXT
TOOL_PURPOSE_MODIFYING_FEATURES_BIT_EXT specifies that the tool modifies the API features/limits/extensions presented to the application.

pattern TOOL_PURPOSE_DEBUG_MARKERS_BIT_EXT :: ToolPurposeFlagBitsEXT
TOOL_PURPOSE_DEBUG_MARKERS_BIT_EXT specifies that the tool consumes debug markers or object debug annotation, queue labels, or command buffer labels

pattern TOOL_PURPOSE_DEBUG_REPORTING_BIT_EXT :: ToolPurposeFlagBitsEXT
TOOL_PURPOSE_DEBUG_REPORTING_BIT_EXT specifies that the tool reports additional information to the application via callbacks specified by createDebugReportCallbackEXT or createDebugUtilsMessengerEXT

type ToolPurposeFlagsEXT = ToolPurposeFlagBitsEXT

type EXT_TOOLING_INFO_SPEC_VERSION = 1

pattern EXT_TOOLING_INFO_SPEC_VERSION :: forall a. Integral a => a

type EXT_TOOLING_INFO_EXTENSION_NAME = "VK_EXT_tooling_info"

pattern EXT_TOOLING_INFO_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a