Vulkan.Extensions.VK_EXT_transform_feedback

Documentation

cmdBindTransformFeedbackBuffersEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command is recorded.

-> ("firstBinding" ::: Word32)
firstBinding is the index of the first transform feedback binding whose state is updated by the command.

-> ("buffers" ::: Vector Buffer)
pBuffers is a pointer to an array of buffer handles.

-> ("offsets" ::: Vector DeviceSize)
pOffsets is a pointer to an array of buffer offsets.

-> ("sizes" ::: Vector DeviceSize)
pSizes is an optional array of buffer sizes, specifying the maximum number of bytes to capture to the corresponding transform feedback buffer. If pSizes is NULL, or the value of the pSizes array element is WHOLE_SIZE, then the maximum bytes captured will be the size of the corresponding buffer minus the buffer offset.

-> io ()

vkCmdBindTransformFeedbackBuffersEXT - Bind transform feedback buffers to a command buffer

Description

The values taken from elements i of pBuffers, pOffsets and pSizes replace the current state for the transform feedback binding firstBinding + i, for i in [0, bindingCount). The transform feedback binding is updated to start at the offset indicated by pOffsets[i] from the start of the buffer pBuffers[i].

Valid Usage

PhysicalDeviceTransformFeedbackFeaturesEXT::transformFeedback must be enabled

firstBinding must be less than PhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackBuffers

The sum of firstBinding and bindingCount must be less than or equal to PhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackBuffers

All elements of pOffsets must be less than the size of the corresponding element in pBuffers

All elements of pOffsets must be a multiple of 4

All elements of pBuffers must have been created with the BUFFER_USAGE_TRANSFORM_FEEDBACK_BUFFER_BIT_EXT flag

If the optional pSize array is specified, each element of pSizes must either be WHOLE_SIZE, or be less than or equal to PhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackBufferSize

All elements of pSizes must be either WHOLE_SIZE, or less than or equal to the size of the corresponding buffer in pBuffers

All elements of pOffsets plus pSizes, where the pSizes, element is not WHOLE_SIZE, must be less than or equal to the size of the corresponding buffer in pBuffers

Each element of pBuffers that is non-sparse must be bound completely and contiguously to a single DeviceMemory object

Transform feedback must not be active when the cmdBindTransformFeedbackBuffersEXT command is recorded

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pBuffers must be a valid pointer to an array of bindingCount valid Buffer handles

pOffsets must be a valid pointer to an array of bindingCount DeviceSize values

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

bindingCount must be greater than 0

Both of commandBuffer, and the elements of pBuffers must have been created, allocated, or retrieved from the same Device

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

Buffer, CommandBuffer, DeviceSize

cmdBeginTransformFeedbackEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command is recorded.

-> ("firstCounterBuffer" ::: Word32)
firstCounterBuffer is the index of the first transform feedback buffer corresponding to pCounterBuffers[0] and pCounterBufferOffsets[0].

-> ("counterBuffers" ::: Vector Buffer)
pCounterBuffers is an optional array of buffer handles to the counter buffers which contain a 4 byte integer value representing the byte offset from the start of the corresponding transform feedback buffer from where to start capturing vertex data. If the byte offset stored to the counter buffer location was done using cmdEndTransformFeedbackEXT it can be used to resume transform feedback from the previous location. If pCounterBuffers is NULL, then transform feedback will start capturing vertex data to byte offset zero in all bound transform feedback buffers. For each element of pCounterBuffers that is NULL_HANDLE, transform feedback will start capturing vertex data to byte zero in the corresponding bound transform feedback buffer.

-> ("counterBufferOffsets" ::: Vector DeviceSize)
pCounterBufferOffsets is an optional array of offsets within each of the pCounterBuffers where the counter values were previously written. The location in each counter buffer at these offsets must be large enough to contain 4 bytes of data. This data is the number of bytes captured by the previous transform feedback to this buffer. If pCounterBufferOffsets is NULL, then it is assumed the offsets are zero.

-> io ()

vkCmdBeginTransformFeedbackEXT - Make transform feedback active in the command buffer

Description

The active transform feedback buffers will capture primitives emitted from the corresponding XfbBuffer in the bound graphics pipeline. Any XfbBuffer emitted that does not output to an active transform feedback buffer will not be captured.

Valid Usage

PhysicalDeviceTransformFeedbackFeaturesEXT::transformFeedback must be enabled

Transform feedback must not be active

firstCounterBuffer must be less than PhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackBuffers

The sum of firstCounterBuffer and counterBufferCount must be less than or equal to PhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackBuffers

If counterBufferCount is not 0, and pCounterBuffers is not NULL, pCounterBuffers must be a valid pointer to an array of counterBufferCount Buffer handles that are either valid or NULL_HANDLE

For each buffer handle in the array, if it is not NULL_HANDLE it must reference a buffer large enough to hold 4 bytes at the corresponding offset from the pCounterBufferOffsets array

If pCounterBuffer is NULL, then pCounterBufferOffsets must also be NULL

For each buffer handle in the pCounterBuffers array that is not NULL_HANDLE it must have been created with a usage value containing BUFFER_USAGE_TRANSFORM_FEEDBACK_COUNTER_BUFFER_BIT_EXT

The last vertex processing stage of the bound graphics pipeline must have been declared with the Xfb execution mode

Transform feedback must not be made active in a render pass instance with multiview enabled

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

If counterBufferCount is not 0, and pCounterBufferOffsets is not NULL, pCounterBufferOffsets must be a valid pointer to an array of counterBufferCount DeviceSize values

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

This command must only be called inside of a render pass instance

Both of commandBuffer, and the elements of pCounterBuffers that are valid handles of non-ignored parameters must have been created, allocated, or retrieved from the same Device

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Inside 	 Graphics

See Also

Buffer, CommandBuffer, DeviceSize

cmdUseTransformFeedbackEXT :: forall io r. MonadIO io => CommandBuffer -> Word32 -> Vector Buffer -> Vector DeviceSize -> io r -> io r

This function will call the supplied action between calls to cmdBeginTransformFeedbackEXT and cmdEndTransformFeedbackEXT

Note that cmdEndTransformFeedbackEXT is *not* called if an exception is thrown by the inner action.

cmdEndTransformFeedbackEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command is recorded.

-> ("firstCounterBuffer" ::: Word32)
firstCounterBuffer is the index of the first transform feedback buffer corresponding to pCounterBuffers[0] and pCounterBufferOffsets[0].

-> ("counterBuffers" ::: Vector Buffer)
pCounterBuffers is an optional array of buffer handles to the counter buffers used to record the current byte positions of each transform feedback buffer where the next vertex output data would be captured. This can be used by a subsequent cmdBeginTransformFeedbackEXT call to resume transform feedback capture from this position. It can also be used by cmdDrawIndirectByteCountEXT to determine the vertex count of the draw call.

-> ("counterBufferOffsets" ::: Vector DeviceSize)
pCounterBufferOffsets is an optional array of offsets within each of the pCounterBuffers where the counter values can be written. The location in each counter buffer at these offsets must be large enough to contain 4 bytes of data. The data stored at this location is the byte offset from the start of the transform feedback buffer binding where the next vertex data would be written. If pCounterBufferOffsets is NULL, then it is assumed the offsets are zero.

-> io ()

vkCmdEndTransformFeedbackEXT - Make transform feedback inactive in the command buffer

Valid Usage

PhysicalDeviceTransformFeedbackFeaturesEXT::transformFeedback must be enabled

Transform feedback must be active

firstCounterBuffer must be less than PhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackBuffers

The sum of firstCounterBuffer and counterBufferCount must be less than or equal to PhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackBuffers

If counterBufferCount is not 0, and pCounterBuffers is not NULL, pCounterBuffers must be a valid pointer to an array of counterBufferCount Buffer handles that are either valid or NULL_HANDLE

For each buffer handle in the array, if it is not NULL_HANDLE it must reference a buffer large enough to hold 4 bytes at the corresponding offset from the pCounterBufferOffsets array

If pCounterBuffer is NULL, then pCounterBufferOffsets must also be NULL

For each buffer handle in the pCounterBuffers array that is not NULL_HANDLE it must have been created with a usage value containing BUFFER_USAGE_TRANSFORM_FEEDBACK_COUNTER_BUFFER_BIT_EXT

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

If counterBufferCount is not 0, and pCounterBufferOffsets is not NULL, pCounterBufferOffsets must be a valid pointer to an array of counterBufferCount DeviceSize values

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

This command must only be called inside of a render pass instance

Both of commandBuffer, and the elements of pCounterBuffers that are valid handles of non-ignored parameters must have been created, allocated, or retrieved from the same Device

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Inside 	 Graphics

See Also

Buffer, CommandBuffer, DeviceSize

cmdBeginQueryIndexedEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which this command will be recorded.

-> QueryPool
queryPool is the query pool that will manage the results of the query.

-> ("query" ::: Word32)
query is the query index within the query pool that will contain the results.

-> QueryControlFlags
flags is a bitmask of QueryControlFlagBits specifying constraints on the types of queries that can be performed.

-> ("index" ::: Word32)
index is the query type specific index. When the query type is QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT the index represents the vertex stream.

-> io ()

vkCmdBeginQueryIndexedEXT - Begin an indexed query

Description

The cmdBeginQueryIndexedEXT command operates the same as the cmdBeginQuery command, except that it also accepts a query type specific index parameter.

Valid Usage

queryPool must have been created with a queryType that differs from that of any queries that are active within commandBuffer

All queries used by the command must be unavailable

The queryType used to create queryPool must not be QUERY_TYPE_TIMESTAMP

If the precise occlusion queries feature is not enabled, or the queryType used to create queryPool was not QUERY_TYPE_OCCLUSION, flags must not contain QUERY_CONTROL_PRECISE_BIT

query must be less than the number of queries in queryPool

If the queryType used to create queryPool was QUERY_TYPE_OCCLUSION, the CommandPool that commandBuffer was allocated from must support graphics operations

If the queryType used to create queryPool was QUERY_TYPE_PIPELINE_STATISTICS and any of the pipelineStatistics indicate graphics operations, the CommandPool that commandBuffer was allocated from must support graphics operations

If the queryType used to create queryPool was QUERY_TYPE_PIPELINE_STATISTICS and any of the pipelineStatistics indicate compute operations, the CommandPool that commandBuffer was allocated from must support compute operations

commandBuffer must not be a protected command buffer

If called within a render pass instance, the sum of query and the number of bits set in the current subpass’s view mask must be less than or equal to the number of queries in queryPool

If the queryType used to create queryPool was QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT the CommandPool that commandBuffer was allocated from must support graphics operations

If the queryType used to create queryPool was QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT the index parameter must be less than PhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackStreams

If the queryType used to create queryPool was not QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT the index must be zero

If the queryType used to create queryPool was QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT then PhysicalDeviceTransformFeedbackPropertiesEXT::transformFeedbackQueries must be supported

If queryPool was created with a queryType of QUERY_TYPE_PERFORMANCE_QUERY_KHR, the profiling lock must have been held before beginCommandBuffer was called on commandBuffer

If queryPool was created with a queryType of QUERY_TYPE_PERFORMANCE_QUERY_KHR and one of the counters used to create queryPool was PERFORMANCE_COUNTER_SCOPE_COMMAND_BUFFER_KHR, the query begin must be the first recorded command in commandBuffer

If queryPool was created with a queryType of QUERY_TYPE_PERFORMANCE_QUERY_KHR and one of the counters used to create queryPool was PERFORMANCE_COUNTER_SCOPE_RENDER_PASS_KHR, the begin command must not be recorded within a render pass instance

If queryPool was created with a queryType of QUERY_TYPE_PERFORMANCE_QUERY_KHR and another query pool with a queryType QUERY_TYPE_PERFORMANCE_QUERY_KHR has been used within commandBuffer, its parent primary command buffer or secondary command buffer recorded within the same parent primary command buffer as commandBuffer, the performanceCounterMultipleQueryPools feature must be enabled

If queryPool was created with a queryType of QUERY_TYPE_PERFORMANCE_QUERY_KHR, this command must not be recorded in a command buffer that, either directly or through secondary command buffers, also contains a cmdResetQueryPool command affecting the same query

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

queryPool must be a valid QueryPool handle

flags must be a valid combination of QueryControlFlagBits values

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, or compute operations

Both of commandBuffer, and queryPool must have been created, allocated, or retrieved from the same Device

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute

See Also

CommandBuffer, QueryControlFlags, QueryPool

cmdUseQueryIndexedEXT :: forall io r. MonadIO io => CommandBuffer -> QueryPool -> Word32 -> QueryControlFlags -> Word32 -> io r -> io r

This function will call the supplied action between calls to cmdBeginQueryIndexedEXT and cmdEndQueryIndexedEXT

Note that cmdEndQueryIndexedEXT is *not* called if an exception is thrown by the inner action.

cmdEndQueryIndexedEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which this command will be recorded.

-> QueryPool
queryPool is the query pool that is managing the results of the query.

-> ("query" ::: Word32)
query is the query index within the query pool where the result is stored.

-> ("index" ::: Word32)
index is the query type specific index.

-> io ()

vkCmdEndQueryIndexedEXT - Ends a query

Description

The cmdEndQueryIndexedEXT command operates the same as the cmdEndQuery command, except that it also accepts a query type specific index parameter.

Valid Usage

All queries used by the command must be active

query must be less than the number of queries in queryPool

commandBuffer must not be a protected command buffer

If cmdEndQueryIndexedEXT is called within a render pass instance, the sum of query and the number of bits set in the current subpass’s view mask must be less than or equal to the number of queries in queryPool

If the queryType used to create queryPool was QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT the index parameter must be less than PhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackStreams

If the queryType used to create queryPool was not QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT the index must be zero

If the queryType used to create queryPool was QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT index must equal the index used to begin the query

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

queryPool must be a valid QueryPool handle

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, or compute operations

Both of commandBuffer, and queryPool must have been created, allocated, or retrieved from the same Device

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute

See Also

CommandBuffer, QueryPool

cmdDrawIndirectByteCountEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command is recorded.

-> ("instanceCount" ::: Word32)
instanceCount is the number of instances to draw.

-> ("firstInstance" ::: Word32)
firstInstance is the instance ID of the first instance to draw.

-> ("counterBuffer" ::: Buffer)
counterBuffer is the buffer handle from where the byte count is read.

-> ("counterBufferOffset" ::: DeviceSize)
counterBufferOffset is the offset into the buffer used to read the byte count, which is used to calculate the vertex count for this draw call.

-> ("counterOffset" ::: Word32)
counterOffset is subtracted from the byte count read from the counterBuffer at the counterBufferOffset

-> ("vertexStride" ::: Word32)
vertexStride is the stride in bytes between each element of the vertex data that is used to calculate the vertex count from the counter value. This value is typically the same value that was used in the graphics pipeline state when the transform feedback was captured as the XfbStride.

-> io ()

vkCmdDrawIndirectByteCountEXT - Draw primitives where the vertex count is derived from the counter byte value in the counter buffer

Description

When the command is executed, primitives are assembled in the same way as done with cmdDraw except the vertexCount is calculated based on the byte count read from counterBuffer at offset counterBufferOffset. The assembled primitives execute the bound graphics pipeline.

The effective vertexCount is calculated as follows:

const uint32_t * counterBufferPtr = (const uint8_t *)counterBuffer.address + counterBufferOffset;
vertexCount = floor(max(0, (*counterBufferPtr - counterOffset)) / vertexStride);

The effective firstVertex is zero.

Valid Usage

If a ImageView is sampled with FILTER_LINEAR as a result of this command, then the image view’s format features must contain FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_LINEAR_BIT

If a ImageView is accessed using atomic operations as a result of this command, then the image view’s format features must contain FORMAT_FEATURE_STORAGE_IMAGE_ATOMIC_BIT

If a ImageView is sampled with FILTER_CUBIC_EXT as a result of this command, then the image view’s format features must contain FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_CUBIC_BIT_EXT

Any ImageView being sampled with FILTER_CUBIC_EXT as a result of this command must have a ImageViewType and format that supports cubic filtering, as specified by FilterCubicImageViewImageFormatPropertiesEXT::filterCubic returned by getPhysicalDeviceImageFormatProperties2

Any ImageView being sampled with FILTER_CUBIC_EXT with a reduction mode of either SAMPLER_REDUCTION_MODE_MIN or SAMPLER_REDUCTION_MODE_MAX as a result of this command must have a ImageViewType and format that supports cubic filtering together with minmax filtering, as specified by FilterCubicImageViewImageFormatPropertiesEXT::filterCubicMinmax returned by getPhysicalDeviceImageFormatProperties2

Any Image created with a ImageCreateInfo::flags containing IMAGE_CREATE_CORNER_SAMPLED_BIT_NV sampled as a result of this command must only be sampled using a SamplerAddressMode of SAMPLER_ADDRESS_MODE_CLAMP_TO_EDGE

For each set n that is statically used by the Pipeline bound to the pipeline bind point used by this command, a descriptor set must have been bound to n at the same pipeline bind point, with a PipelineLayout that is compatible for set n, with the PipelineLayout used to create the current Pipeline, as described in ???

For each push constant that is statically used by the Pipeline bound to the pipeline bind point used by this command, a push constant value must have been set for the same pipeline bind point, with a PipelineLayout that is compatible for push constants, with the PipelineLayout used to create the current Pipeline, as described in ???

Descriptors in each bound descriptor set, specified via cmdBindDescriptorSets, must be valid if they are statically used by the Pipeline bound to the pipeline bind point used by this command

A valid pipeline must be bound to the pipeline bind point used by this command

If the Pipeline object bound to the pipeline bind point used by this command requires any dynamic state, that state must have been set for commandBuffer, and done so after any previously bound pipeline with the corresponding state not specified as dynamic

There must not have been any calls to dynamic state setting commands for any state not specified as dynamic in the Pipeline object bound to the pipeline bind point used by this command, since that pipeline was bound

If the Pipeline object bound to the pipeline bind point used by this command accesses a Sampler object that uses unnormalized coordinates, that sampler must not be used to sample from any Image with a ImageView of the type IMAGE_VIEW_TYPE_3D, IMAGE_VIEW_TYPE_CUBE, IMAGE_VIEW_TYPE_1D_ARRAY, IMAGE_VIEW_TYPE_2D_ARRAY or IMAGE_VIEW_TYPE_CUBE_ARRAY, in any shader stage

If the Pipeline object bound to the pipeline bind point used by this command accesses a Sampler object that uses unnormalized coordinates, that sampler must not be used with any of the SPIR-V OpImageSample* or OpImageSparseSample* instructions with ImplicitLod, Dref or Proj in their name, in any shader stage

If the Pipeline object bound to the pipeline bind point used by this command accesses a Sampler object that uses unnormalized coordinates, that sampler must not be used with any of the SPIR-V OpImageSample* or OpImageSparseSample* instructions that includes a LOD bias or any offset values, in any shader stage

If the robust buffer access feature is not enabled, and if the Pipeline object bound to the pipeline bind point used by this command accesses a uniform buffer, it must not access values outside of the range of the buffer as specified in the descriptor set bound to the same pipeline bind point

If the robust buffer access feature is not enabled, and if the Pipeline object bound to the pipeline bind point used by this command accesses a storage buffer, it must not access values outside of the range of the buffer as specified in the descriptor set bound to the same pipeline bind point

If commandBuffer is an unprotected command buffer, any resource accessed by the Pipeline object bound to the pipeline bind point used by this command must not be a protected resource

If a ImageView is accessed using OpImageWrite as a result of this command, then the Type of the Texel operand of that instruction must have at least as many components as the image view’s format.

If a BufferView is accessed using OpImageWrite as a result of this command, then the Type of the Texel operand of that instruction must have at least as many components as the image view’s format.

If a ImageView with a Format that has a 64-bit channel width is accessed as a result of this command, the SampledType of the OpTypeImage operand of that instruction must have a Width of 64.

If a ImageView with a Format that has a channel width less than 64-bit is accessed as a result of this command, the SampledType of the OpTypeImage operand of that instruction must have a Width of 32.

If a BufferView with a Format that has a 64-bit channel width is accessed as a result of this command, the SampledType of the OpTypeImage operand of that instruction must have a Width of 64.

If a BufferView with a Format that has a channel width less than 64-bit is accessed as a result of this command, the SampledType of the OpTypeImage operand of that instruction must have a Width of 32.

If the sparseImageInt64Atomics feature is not enabled, Image objects created with the IMAGE_CREATE_SPARSE_RESIDENCY_BIT flag must not be accessed by atomic instructions through an OpTypeImage with a SampledType with a Width of 64 by this command.

If the sparseImageInt64Atomics feature is not enabled, Buffer objects created with the BUFFER_CREATE_SPARSE_RESIDENCY_BIT flag must not be accessed by atomic instructions through an OpTypeImage with a SampledType with a Width of 64 by this command.

The current render pass must be compatible with the renderPass member of the GraphicsPipelineCreateInfo structure specified when creating the Pipeline bound to PIPELINE_BIND_POINT_GRAPHICS

The subpass index of the current render pass must be equal to the subpass member of the GraphicsPipelineCreateInfo structure specified when creating the Pipeline bound to PIPELINE_BIND_POINT_GRAPHICS

Every input attachment used by the current subpass must be bound to the pipeline via a descriptor set

Image subresources used as attachments in the current render pass must not be accessed in any way other than as an attachment by this command

If the draw is recorded in a render pass instance with multiview enabled, the maximum instance index must be less than or equal to PhysicalDeviceMultiviewProperties::maxMultiviewInstanceIndex

If the bound graphics pipeline was created with PipelineSampleLocationsStateCreateInfoEXT::sampleLocationsEnable set to TRUE and the current subpass has a depth/stencil attachment, then that attachment must have been created with the IMAGE_CREATE_SAMPLE_LOCATIONS_COMPATIBLE_DEPTH_BIT_EXT bit set

If the bound graphics pipeline state was created with the DYNAMIC_STATE_VIEWPORT_WITH_COUNT_EXT dynamic state enabled, but not the DYNAMIC_STATE_SCISSOR_WITH_COUNT_EXT dynamic state enabled, then cmdSetViewportWithCountEXT must have been called in the current command buffer prior to this draw command, and the viewportCount parameter of cmdSetViewportWithCountEXT must match the PipelineViewportStateCreateInfo::scissorCount of the pipeline

If the bound graphics pipeline state was created with the DYNAMIC_STATE_SCISSOR_WITH_COUNT_EXT dynamic state enabled, but not the DYNAMIC_STATE_VIEWPORT_WITH_COUNT_EXT dynamic state enabled, then cmdSetScissorWithCountEXT must have been called in the current command buffer prior to this draw command, and the scissorCount parameter of cmdSetScissorWithCountEXT must match the PipelineViewportStateCreateInfo::viewportCount of the pipeline

If the bound graphics pipeline state was created with both the DYNAMIC_STATE_SCISSOR_WITH_COUNT_EXT and DYNAMIC_STATE_VIEWPORT_WITH_COUNT_EXT dynamic states enabled then both cmdSetViewportWithCountEXT and cmdSetScissorWithCountEXT must have been called in the current command buffer prior to this draw command, and the viewportCount parameter of cmdSetViewportWithCountEXT must match the scissorCount parameter of cmdSetScissorWithCountEXT

If the bound graphics pipeline state was created with the DYNAMIC_STATE_VIEWPORT_WITH_COUNT_EXT dynamic state enabled, but not the DYNAMIC_STATE_VIEWPORT_W_SCALING_NV dynamic state enabled, then the bound graphics pipeline must have been created with PipelineViewportWScalingStateCreateInfoNV::viewportCount greater or equal to the viewportCount parameter in the last call to cmdSetViewportWithCountEXT

If the bound graphics pipeline state was created with the DYNAMIC_STATE_VIEWPORT_WITH_COUNT_EXT and DYNAMIC_STATE_VIEWPORT_W_SCALING_NV dynamic states enabled then the viewportCount parameter in the last call to cmdSetViewportWScalingNV must be greater than or equal to the viewportCount parameter in the last call to cmdSetViewportWithCountEXT

If the bound graphics pipeline state was created with the DYNAMIC_STATE_VIEWPORT_WITH_COUNT_EXT dynamic state enabled, but not the DYNAMIC_STATE_VIEWPORT_SHADING_RATE_PALETTE_NV dynamic state enabled, then the bound graphics pipeline must have been created with PipelineViewportShadingRateImageStateCreateInfoNV::viewportCount greater or equal to the viewportCount parameter in the last call to cmdSetViewportWithCountEXT

If the bound graphics pipeline state was created with the DYNAMIC_STATE_VIEWPORT_WITH_COUNT_EXT and DYNAMIC_STATE_VIEWPORT_SHADING_RATE_PALETTE_NV dynamic states enabled then the viewportCount parameter in the last call to cmdSetViewportShadingRatePaletteNV must be greater than or equal to the viewportCount parameter in the last call to cmdSetViewportWithCountEXT

If the bound graphics pipeline state was created with the DYNAMIC_STATE_VIEWPORT_WITH_COUNT_EXT dynamic state enabled and an instance of PipelineViewportSwizzleStateCreateInfoNV chained from VkPipelineVieportCreateInfo, then the bound graphics pipeline must have been created with PipelineViewportSwizzleStateCreateInfoNV::viewportCount greater or equal to the viewportCount parameter in the last call to cmdSetViewportWithCountEXT

If the bound graphics pipeline state was created with the DYNAMIC_STATE_VIEWPORT_WITH_COUNT_EXT dynamic state enabled and an instance of PipelineViewportExclusiveScissorStateCreateInfoNV chained from VkPipelineVieportCreateInfo, then the bound graphics pipeline must have been created with PipelineViewportExclusiveScissorStateCreateInfoNV::exclusiveScissorCount greater or equal to the viewportCount parameter in the last call to cmdSetViewportWithCountEXT

If the bound graphics pipeline state was created with the DYNAMIC_STATE_PRIMITIVE_TOPOLOGY_EXT dynamic state enabled then cmdSetPrimitiveTopologyEXT must have been called in the current command buffer prior to this draw command, and the primitiveTopology parameter of cmdSetPrimitiveTopologyEXT must be of the same topology class as the pipeline PipelineInputAssemblyStateCreateInfo::topology state

All vertex input bindings accessed via vertex input variables declared in the vertex shader entry point’s interface must have either valid or NULL_HANDLE buffers bound

If the nullDescriptor feature is not enabled, all vertex input bindings accessed via vertex input variables declared in the vertex shader entry point’s interface must not be NULL_HANDLE

For a given vertex buffer binding, any attribute data fetched must be entirely contained within the corresponding vertex buffer binding, as described in ???

PhysicalDeviceTransformFeedbackFeaturesEXT::transformFeedback must be enabled

The implementation must support PhysicalDeviceTransformFeedbackPropertiesEXT::transformFeedbackDraw

vertexStride must be greater than 0 and less than or equal to PhysicalDeviceLimits::maxTransformFeedbackBufferDataStride

counterBuffer must have been created with the BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set

commandBuffer must not be a protected command buffer

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

counterBuffer must be a valid Buffer handle

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

This command must only be called inside of a render pass instance

Both of commandBuffer, and counterBuffer must have been created, allocated, or retrieved from the same Device

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Inside 	 Graphics 	 Graphics

See Also

Buffer, CommandBuffer, DeviceSize

data PhysicalDeviceTransformFeedbackFeaturesEXT

VkPhysicalDeviceTransformFeedbackFeaturesEXT - Structure describing transform feedback features that can be supported by an implementation

Members

The members of the PhysicalDeviceTransformFeedbackFeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceTransformFeedbackFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether each feature is supported. PhysicalDeviceTransformFeedbackFeaturesEXT can also be included in the pNext chain of DeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceTransformFeedbackFeaturesEXT

Fields

transformFeedback :: Bool

transformFeedback indicates whether the implementation supports transform feedback and shader modules can declare the TransformFeedback capability.

geometryStreams :: Bool

geometryStreams indicates whether the implementation supports the GeometryStreams SPIR-V capability.

data PhysicalDeviceTransformFeedbackPropertiesEXT

VkPhysicalDeviceTransformFeedbackPropertiesEXT - Structure describing transform feedback properties that can be supported by an implementation

Members

The members of the PhysicalDeviceTransformFeedbackPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the PhysicalDeviceTransformFeedbackPropertiesEXT structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits and properties.

Valid Usage (Implicit)

See Also

Bool32, DeviceSize, StructureType

Constructors

PhysicalDeviceTransformFeedbackPropertiesEXT

Fields

maxTransformFeedbackStreams :: Word32

maxTransformFeedbackStreams is the maximum number of vertex streams that can be output from geometry shaders declared with the GeometryStreams capability. If the implementation does not support PhysicalDeviceTransformFeedbackFeaturesEXT::geometryStreams then maxTransformFeedbackStreams must be set to 1.

maxTransformFeedbackBuffers :: Word32

maxTransformFeedbackBuffers is the maximum number of transform feedback buffers that can be bound for capturing shader outputs from the last vertex processing stage.

maxTransformFeedbackBufferSize :: DeviceSize

maxTransformFeedbackBufferSize is the maximum size that can be specified when binding a buffer for transform feedback in cmdBindTransformFeedbackBuffersEXT.

maxTransformFeedbackStreamDataSize :: Word32

maxTransformFeedbackStreamDataSize is the maximum amount of data in bytes for each vertex that captured to one or more transform feedback buffers associated with a specific vertex stream.

maxTransformFeedbackBufferDataSize :: Word32

maxTransformFeedbackBufferDataSize is the maximum amount of data in bytes for each vertex that can be captured to a specific transform feedback buffer.

maxTransformFeedbackBufferDataStride :: Word32

maxTransformFeedbackBufferDataStride is the maximum stride between each capture of vertex data to the buffer.

transformFeedbackQueries :: Bool

transformFeedbackQueries is true if the implementation supports the QUERY_TYPE_TRANSFORM_FEEDBACK_STREAM_EXT query type. transformFeedbackQueries is false if queries of this type cannot be created.

transformFeedbackStreamsLinesTriangles :: Bool

transformFeedbackStreamsLinesTriangles is true if the implementation supports the geometry shader OpExecutionMode of OutputLineStrip and OutputTriangleStrip in addition to OutputPoints when more than one vertex stream is output. If transformFeedbackStreamsLinesTriangles is false the implementation only supports an OpExecutionMode of OutputPoints when more than one vertex stream is output from the geometry shader.

transformFeedbackRasterizationStreamSelect :: Bool

transformFeedbackRasterizationStreamSelect is true if the implementation supports the GeometryStreams SPIR-V capability and the application can use PipelineRasterizationStateStreamCreateInfoEXT to modify which vertex stream output is used for rasterization. Otherwise vertex stream 0 must always be used for rasterization.

transformFeedbackDraw :: Bool

transformFeedbackDraw is true if the implementation supports the cmdDrawIndirectByteCountEXT function otherwise the function must not be called.

data PipelineRasterizationStateStreamCreateInfoEXT

VkPipelineRasterizationStateStreamCreateInfoEXT - Structure defining the geometry stream used for rasterization

Description

If this structure is not present, rasterizationStream is assumed to be zero.

Valid Usage (Implicit)

See Also

PipelineRasterizationStateStreamCreateFlagsEXT, StructureType

Constructors

PipelineRasterizationStateStreamCreateInfoEXT

Fields

flags :: PipelineRasterizationStateStreamCreateFlagsEXT

flags is reserved for future use.

flags must be 0

rasterizationStream :: Word32

rasterizationStream is the vertex stream selected for rasterization.

rasterizationStream must be less than PhysicalDeviceTransformFeedbackPropertiesEXT::maxTransformFeedbackStreams

rasterizationStream must be zero if PhysicalDeviceTransformFeedbackPropertiesEXT::transformFeedbackRasterizationStreamSelect is FALSE

newtype PipelineRasterizationStateStreamCreateFlagsEXT

VkPipelineRasterizationStateStreamCreateFlagsEXT - Reserved for future use

Description

PipelineRasterizationStateStreamCreateFlagsEXT is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

PipelineRasterizationStateStreamCreateInfoEXT

Constructors

PipelineRasterizationStateStreamCreateFlagsEXT Flags

type EXT_TRANSFORM_FEEDBACK_SPEC_VERSION = 1

pattern EXT_TRANSFORM_FEEDBACK_SPEC_VERSION :: forall a. Integral a => a

type EXT_TRANSFORM_FEEDBACK_EXTENSION_NAME = "VK_EXT_transform_feedback"

pattern EXT_TRANSFORM_FEEDBACK_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a