Vulkan.Extensions.VK_EXT_validation_features

Documentation

data ValidationFeaturesEXT

VkValidationFeaturesEXT - Specify validation features to enable or disable for a Vulkan instance

Valid Usage

If the pEnabledValidationFeatures array contains VALIDATION_FEATURE_ENABLE_GPU_ASSISTED_RESERVE_BINDING_SLOT_EXT, then it must also contain VALIDATION_FEATURE_ENABLE_GPU_ASSISTED_EXT

If the pEnabledValidationFeatures array contains VALIDATION_FEATURE_ENABLE_DEBUG_PRINTF_EXT, then it must not contain VALIDATION_FEATURE_ENABLE_GPU_ASSISTED_EXT

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_VALIDATION_FEATURES_EXT

If enabledValidationFeatureCount is not 0, pEnabledValidationFeatures must be a valid pointer to an array of enabledValidationFeatureCount valid ValidationFeatureEnableEXT values

If disabledValidationFeatureCount is not 0, pDisabledValidationFeatures must be a valid pointer to an array of disabledValidationFeatureCount valid ValidationFeatureDisableEXT values

See Also

StructureType, ValidationFeatureDisableEXT, ValidationFeatureEnableEXT

Constructors

ValidationFeaturesEXT

Fields

enabledValidationFeatures :: Vector ValidationFeatureEnableEXT

pEnabledValidationFeatures is a pointer to an array of ValidationFeatureEnableEXT values specifying the validation features to be enabled.

disabledValidationFeatures :: Vector ValidationFeatureDisableEXT

pDisabledValidationFeatures is a pointer to an array of ValidationFeatureDisableEXT values specifying the validation features to be disabled.

newtype ValidationFeatureEnableEXT

VkValidationFeatureEnableEXT - Specify validation features to enable

See Also

ValidationFeaturesEXT

Constructors

ValidationFeatureEnableEXT Int32

Bundled Patterns

pattern VALIDATION_FEATURE_ENABLE_GPU_ASSISTED_EXT :: ValidationFeatureEnableEXT
VALIDATION_FEATURE_ENABLE_GPU_ASSISTED_EXT specifies that GPU-assisted validation is enabled. Activating this feature instruments shader programs to generate additional diagnostic data. This feature is disabled by default.

pattern VALIDATION_FEATURE_ENABLE_GPU_ASSISTED_RESERVE_BINDING_SLOT_EXT :: ValidationFeatureEnableEXT
VALIDATION_FEATURE_ENABLE_GPU_ASSISTED_RESERVE_BINDING_SLOT_EXT specifies that the validation layers reserve a descriptor set binding slot for their own use. The layer reports a value for PhysicalDeviceLimits::maxBoundDescriptorSets that is one less than the value reported by the device. If the device supports the binding of only one descriptor set, the validation layer does not perform GPU-assisted validation. This feature is disabled by default.

pattern VALIDATION_FEATURE_ENABLE_BEST_PRACTICES_EXT :: ValidationFeatureEnableEXT
VALIDATION_FEATURE_ENABLE_BEST_PRACTICES_EXT specifies that Vulkan best-practices validation is enabled. Activating this feature enables the output of warnings related to common misuse of the API, but which are not explicitly prohibited by the specification. This feature is disabled by default.

pattern VALIDATION_FEATURE_ENABLE_DEBUG_PRINTF_EXT :: ValidationFeatureEnableEXT
VALIDATION_FEATURE_ENABLE_DEBUG_PRINTF_EXT specifies that the layers will process debugPrintfEXT operations in shaders and send the resulting output to the debug callback. This feature is disabled by default.

pattern VALIDATION_FEATURE_ENABLE_SYNCHRONIZATION_VALIDATION_EXT :: ValidationFeatureEnableEXT
VALIDATION_FEATURE_ENABLE_SYNCHRONIZATION_VALIDATION_EXT specifies that Vulkan synchronization validation is enabled. This feature reports resource access conflicts due to missing or incorrect synchronization operations between actions (Draw, Copy, Dispatch, Blit) reading or writing the same regions of memory. This feature is disabled by default.

newtype ValidationFeatureDisableEXT

VkValidationFeatureDisableEXT - Specify validation features to disable

See Also

ValidationFeaturesEXT

Constructors

ValidationFeatureDisableEXT Int32

Bundled Patterns

pattern VALIDATION_FEATURE_DISABLE_ALL_EXT :: ValidationFeatureDisableEXT
VALIDATION_FEATURE_DISABLE_ALL_EXT specifies that all validation checks are disabled.

pattern VALIDATION_FEATURE_DISABLE_SHADERS_EXT :: ValidationFeatureDisableEXT
VALIDATION_FEATURE_DISABLE_SHADERS_EXT specifies that shader validation is disabled. This feature is enabled by default.

pattern VALIDATION_FEATURE_DISABLE_THREAD_SAFETY_EXT :: ValidationFeatureDisableEXT
VALIDATION_FEATURE_DISABLE_THREAD_SAFETY_EXT specifies that thread safety validation is disabled. This feature is enabled by default.

pattern VALIDATION_FEATURE_DISABLE_API_PARAMETERS_EXT :: ValidationFeatureDisableEXT
VALIDATION_FEATURE_DISABLE_API_PARAMETERS_EXT specifies that stateless parameter validation is disabled. This feature is enabled by default.

pattern VALIDATION_FEATURE_DISABLE_OBJECT_LIFETIMES_EXT :: ValidationFeatureDisableEXT
VALIDATION_FEATURE_DISABLE_OBJECT_LIFETIMES_EXT specifies that object lifetime validation is disabled. This feature is enabled by default.

pattern VALIDATION_FEATURE_DISABLE_CORE_CHECKS_EXT :: ValidationFeatureDisableEXT
VALIDATION_FEATURE_DISABLE_CORE_CHECKS_EXT specifies that core validation checks are disabled. This feature is enabled by default. If this feature is disabled, the shader validation and GPU-assisted validation features are also disabled.

pattern VALIDATION_FEATURE_DISABLE_UNIQUE_HANDLES_EXT :: ValidationFeatureDisableEXT
VALIDATION_FEATURE_DISABLE_UNIQUE_HANDLES_EXT specifies that protection against duplicate non-dispatchable object handles is disabled. This feature is enabled by default.

type EXT_VALIDATION_FEATURES_SPEC_VERSION = 4

pattern EXT_VALIDATION_FEATURES_SPEC_VERSION :: forall a. Integral a => a

type EXT_VALIDATION_FEATURES_EXTENSION_NAME = "VK_EXT_validation_features"

pattern EXT_VALIDATION_FEATURES_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a