Vulkan.Extensions.VK_EXT_validation_flags

Documentation

data ValidationFlagsEXT

VkValidationFlagsEXT - Specify validation checks to disable for a Vulkan instance

Valid Usage (Implicit)

See Also

StructureType, ValidationCheckEXT

Constructors

ValidationFlagsEXT

Fields

disabledValidationChecks :: Vector ValidationCheckEXT

pDisabledValidationChecks must be a valid pointer to an array of disabledValidationCheckCount valid ValidationCheckEXT values

newtype ValidationCheckEXT

VkValidationCheckEXT - Specify validation checks to disable

See Also

ValidationFlagsEXT

Constructors

ValidationCheckEXT Int32

Bundled Patterns

pattern VALIDATION_CHECK_ALL_EXT :: ValidationCheckEXT
VALIDATION_CHECK_ALL_EXT specifies that all validation checks are disabled.

pattern VALIDATION_CHECK_SHADERS_EXT :: ValidationCheckEXT
VALIDATION_CHECK_SHADERS_EXT specifies that shader validation is disabled.

type EXT_VALIDATION_FLAGS_SPEC_VERSION = 2

pattern EXT_VALIDATION_FLAGS_SPEC_VERSION :: forall a. Integral a => a

type EXT_VALIDATION_FLAGS_EXTENSION_NAME = "VK_EXT_validation_flags"

pattern EXT_VALIDATION_FLAGS_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a