Vulkan.Extensions.VK_GGP_frame_token

Description

Name

VK_GGP_frame_token - device extension

VK_GGP_frame_token

Name String

VK_GGP_frame_token

Extension Type

Device extension

Registered Extension Number

192

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_swapchain

Requires VK_GGP_stream_descriptor_surface

Contact

Jean-Francois Roy https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_GGP_frame_token:%20&body=@jfroy%20

Other Extension Metadata

Last Modified Date

2019-01-28

IP Status

No known IP claims.

Contributors

Jean-Francois Roy, Google

Richard O’Grady, Google

Description

This extension allows an application that uses the VK_KHR_swapchain extension in combination with a Google Games Platform surface provided by the VK_GGP_stream_descriptor_surface extension to associate a Google Games Platform frame token with a present operation.

New Structures

Extending PresentInfoKHR:

PresentFrameTokenGGP

New Enum Constants

GGP_FRAME_TOKEN_EXTENSION_NAME

GGP_FRAME_TOKEN_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_PRESENT_FRAME_TOKEN_GGP

Version History

Revision 1, 2018-11-26 (Jean-Francois Roy)

Initial revision.

See Also

PresentFrameTokenGGP

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data PresentFrameTokenGGP

VkPresentFrameTokenGGP - The Google Games Platform frame token

Valid Usage (Implicit)

See Also

StructureType

Constructors

PresentFrameTokenGGP

Fields

frameToken :: GgpFrameToken

frameToken is the Google Games Platform frame token.

frameToken must be a valid GgpFrameToken

type GGP_FRAME_TOKEN_SPEC_VERSION = 1

pattern GGP_FRAME_TOKEN_SPEC_VERSION :: forall a. Integral a => a

type GGP_FRAME_TOKEN_EXTENSION_NAME = "VK_GGP_frame_token"

pattern GGP_FRAME_TOKEN_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

type GgpFrameToken = Word32