Vulkan.Extensions.VK_GOOGLE_display_timing

Documentation

getRefreshCycleDurationGOOGLE

Arguments

:: forall io. MonadIO io

=> Device
device is the device associated with swapchain.

-> SwapchainKHR
swapchain is the swapchain to obtain the refresh duration for.

-> io ("displayTimingProperties" ::: RefreshCycleDurationGOOGLE)

vkGetRefreshCycleDurationGOOGLE - Obtain the RC duration of the PE’s display

Valid Usage (Implicit)

device must be a valid Device handle

swapchain must be a valid SwapchainKHR handle

pDisplayTimingProperties must be a valid pointer to a RefreshCycleDurationGOOGLE structure

Both of device, and swapchain must have been created, allocated, or retrieved from the same Instance

Host Synchronization

Host access to swapchain must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_DEVICE_LOST

ERROR_SURFACE_LOST_KHR

See Also

Device, RefreshCycleDurationGOOGLE, SwapchainKHR

getPastPresentationTimingGOOGLE

Arguments

:: forall io. MonadIO io

=> Device
device is the device associated with swapchain.

-> SwapchainKHR
swapchain is the swapchain to obtain presentation timing information duration for.

-> io (Result, "presentationTimings" ::: Vector PastPresentationTimingGOOGLE)

vkGetPastPresentationTimingGOOGLE - Obtain timing of a previously-presented image

Description

If pPresentationTimings is NULL, then the number of newly-available timing records for the given swapchain is returned in pPresentationTimingCount. Otherwise, pPresentationTimingCount must point to a variable set by the user to the number of elements in the pPresentationTimings array, and on return the variable is overwritten with the number of structures actually written to pPresentationTimings. If the value of pPresentationTimingCount is less than the number of newly-available timing records, at most pPresentationTimingCount structures will be written. If pPresentationTimingCount is smaller than the number of newly-available timing records for the given swapchain, INCOMPLETE will be returned instead of SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

device must be a valid Device handle

swapchain must be a valid SwapchainKHR handle

pPresentationTimingCount must be a valid pointer to a uint32_t value

If the value referenced by pPresentationTimingCount is not 0, and pPresentationTimings is not NULL, pPresentationTimings must be a valid pointer to an array of pPresentationTimingCount PastPresentationTimingGOOGLE structures

Both of device, and swapchain must have been created, allocated, or retrieved from the same Instance

Host Synchronization

Host access to swapchain must be externally synchronized

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_DEVICE_LOST

ERROR_OUT_OF_DATE_KHR

ERROR_SURFACE_LOST_KHR

See Also

Device, PastPresentationTimingGOOGLE, SwapchainKHR

data RefreshCycleDurationGOOGLE

VkRefreshCycleDurationGOOGLE - Structure containing the RC duration of a display

See Also

getRefreshCycleDurationGOOGLE

Constructors

RefreshCycleDurationGOOGLE

Fields

refreshDuration :: Word64

refreshDuration is the number of nanoseconds from the start of one refresh cycle to the next.

data PastPresentationTimingGOOGLE

VkPastPresentationTimingGOOGLE - Structure containing timing information about a previously-presented image

Description

The results for a given swapchain and presentID are only returned once from getPastPresentationTimingGOOGLE.

The application can use the PastPresentationTimingGOOGLE values to occasionally adjust its timing. For example, if actualPresentTime is later than expected (e.g. one refreshDuration late), the application may increase its target IPD to a higher multiple of refreshDuration (e.g. decrease its frame rate from 60Hz to 30Hz). If actualPresentTime and earliestPresentTime are consistently different, and if presentMargin is consistently large enough, the application may decrease its target IPD to a smaller multiple of refreshDuration (e.g. increase its frame rate from 30Hz to 60Hz). If actualPresentTime and earliestPresentTime are same, and if presentMargin is consistently high, the application may delay the start of its input-render-present loop in order to decrease the latency between user input and the corresponding present (always leaving some margin in case a new image takes longer to render than the previous image). An application that desires its target IPD to always be the same as refreshDuration, can also adjust features until actualPresentTime is never late and presentMargin is satisfactory.

See Also

getPastPresentationTimingGOOGLE

Constructors

PastPresentationTimingGOOGLE

Fields

presentID :: Word32

presentID is an application-provided value that was given to a previous queuePresentKHR command via PresentTimeGOOGLE::presentID (see below). It can be used to uniquely identify a previous present with the queuePresentKHR command.

desiredPresentTime :: Word64

desiredPresentTime is an application-provided value that was given to a previous queuePresentKHR command via PresentTimeGOOGLE::desiredPresentTime. If non-zero, it was used by the application to indicate that an image not be presented any sooner than desiredPresentTime.

actualPresentTime :: Word64

actualPresentTime is the time when the image of the swapchain was actually displayed.

earliestPresentTime :: Word64

earliestPresentTime is the time when the image of the swapchain could have been displayed. This may differ from actualPresentTime if the application requested that the image be presented no sooner than PresentTimeGOOGLE::desiredPresentTime.

presentMargin :: Word64

presentMargin is an indication of how early the queuePresentKHR command was processed compared to how soon it needed to be processed, and still be presented at earliestPresentTime.

data PresentTimesInfoGOOGLE

VkPresentTimesInfoGOOGLE - The earliest time each image should be presented

Valid Usage

swapchainCount must be the same value as PresentInfoKHR::swapchainCount, where PresentInfoKHR is included in the pNext chain of this PresentTimesInfoGOOGLE structure

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PRESENT_TIMES_INFO_GOOGLE

If pTimes is not NULL, pTimes must be a valid pointer to an array of swapchainCount PresentTimeGOOGLE structures

swapchainCount must be greater than 0

See Also

PresentTimeGOOGLE, StructureType

Constructors

PresentTimesInfoGOOGLE

Fields

swapchainCount :: Word32

swapchainCount is the number of swapchains being presented to by this command.

times :: Vector PresentTimeGOOGLE

pTimes is NULL or a pointer to an array of PresentTimeGOOGLE elements with swapchainCount entries. If not NULL, each element of pTimes contains the earliest time to present the image corresponding to the entry in the PresentInfoKHR::pImageIndices array.

data PresentTimeGOOGLE

VkPresentTimeGOOGLE - The earliest time image should be presented

See Also

PresentTimesInfoGOOGLE

Constructors

PresentTimeGOOGLE

Fields

presentID :: Word32

presentID is an application-provided identification value, that can be used with the results of getPastPresentationTimingGOOGLE, in order to uniquely identify this present. In order to be useful to the application, it should be unique within some period of time that is meaningful to the application.

desiredPresentTime :: Word64

desiredPresentTime specifies that the image given should not be displayed to the user any earlier than this time. desiredPresentTime is a time in nanoseconds, relative to a monotonically-increasing clock (e.g. CLOCK_MONOTONIC (see clock_gettime(2)) on Android and Linux). A value of zero specifies that the presentation engine may display the image at any time. This is useful when the application desires to provide presentID, but does not need a specific desiredPresentTime.

type GOOGLE_DISPLAY_TIMING_SPEC_VERSION = 1

pattern GOOGLE_DISPLAY_TIMING_SPEC_VERSION :: forall a. Integral a => a

type GOOGLE_DISPLAY_TIMING_EXTENSION_NAME = "VK_GOOGLE_display_timing"

pattern GOOGLE_DISPLAY_TIMING_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype SwapchainKHR

VkSwapchainKHR - Opaque handle to a swapchain object

Description

A swapchain is an abstraction for an array of presentable images that are associated with a surface. The presentable images are represented by Image objects created by the platform. One image (which can be an array image for multiview/stereoscopic-3D surfaces) is displayed at a time, but multiple images can be queued for presentation. An application renders to the image, and then queues the image for presentation to the surface.

A native window cannot be associated with more than one non-retired swapchain at a time. Further, swapchains cannot be created for native windows that have a non-Vulkan graphics API surface associated with them.

Note

The presentation engine is an abstraction for the platform’s compositor or display engine.

The presentation engine may be synchronous or asynchronous with respect to the application and/or logical device.

Some implementations may use the device’s graphics queue or dedicated presentation hardware to perform presentation.

The presentable images of a swapchain are owned by the presentation engine. An application can acquire use of a presentable image from the presentation engine. Use of a presentable image must occur only after the image is returned by acquireNextImageKHR, and before it is presented by queuePresentKHR. This includes transitioning the image layout and rendering commands.

An application can acquire use of a presentable image with acquireNextImageKHR. After acquiring a presentable image and before modifying it, the application must use a synchronization primitive to ensure that the presentation engine has finished reading from the image. The application can then transition the image’s layout, queue rendering commands to it, etc. Finally, the application presents the image with queuePresentKHR, which releases the acquisition of the image.

The presentation engine controls the order in which presentable images are acquired for use by the application.

Note

This allows the platform to handle situations which require out-of-order return of images after presentation. At the same time, it allows the application to generate command buffers referencing all of the images in the swapchain at initialization time, rather than in its main loop.

See Also

AcquireNextImageInfoKHR, BindImageMemorySwapchainInfoKHR, ImageSwapchainCreateInfoKHR, PresentInfoKHR, SwapchainCreateInfoKHR, acquireFullScreenExclusiveModeEXT, acquireNextImageKHR, createSharedSwapchainsKHR, createSwapchainKHR, destroySwapchainKHR, getPastPresentationTimingGOOGLE, getRefreshCycleDurationGOOGLE, getSwapchainCounterEXT, getSwapchainImagesKHR, getSwapchainStatusKHR, releaseFullScreenExclusiveModeEXT, setHdrMetadataEXT, setLocalDimmingAMD

Constructors

SwapchainKHR Word64