Vulkan.Extensions.VK_IMG_filter_cubic

Description

Name

VK_IMG_filter_cubic - device extension

VK_IMG_filter_cubic

Name String

VK_IMG_filter_cubic

Extension Type

Device extension

Registered Extension Number

16

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Contact

Tobias Hector https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_IMG_filter_cubic:%20&body=@tobski%20

Other Extension Metadata

Last Modified Date

2016-02-23

Contributors

Tobias Hector, Imagination Technologies

Description

VK_IMG_filter_cubic adds an additional, high quality cubic filtering mode to Vulkan, using a Catmull-Rom bicubic filter. Performing this kind of filtering can be done in a shader by using 16 samples and a number of instructions, but this can be inefficient. The cubic filter mode exposes an optimized high quality texture sampling using fixed texture sampling functionality.

New Enum Constants

IMG_FILTER_CUBIC_EXTENSION_NAME

IMG_FILTER_CUBIC_SPEC_VERSION

Extending Filter:

FILTER_CUBIC_IMG

Extending FormatFeatureFlagBits:

FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_CUBIC_BIT_IMG

Example

Creating a sampler with the new filter for both magnification and minification

    VkSamplerCreateInfo createInfo =
    {
        VK_STRUCTURE_TYPE_SAMPLER_CREATE_INFO // sType
        // Other members set to application-desired values
    };

    createInfo.magFilter = VK_FILTER_CUBIC_IMG;
    createInfo.minFilter = VK_FILTER_CUBIC_IMG;

    VkSampler sampler;
    VkResult result = vkCreateSampler(
        device,
        &createInfo,
        &sampler);

Version History

Revision 1, 2016-02-23 (Tobias Hector)

Initial version

See Also

No cross-references are available

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

type IMG_FILTER_CUBIC_SPEC_VERSION = 1

pattern IMG_FILTER_CUBIC_SPEC_VERSION :: forall a. Integral a => a

type IMG_FILTER_CUBIC_EXTENSION_NAME = "VK_IMG_filter_cubic"

pattern IMG_FILTER_CUBIC_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a