Vulkan.Extensions.VK_IMG_format_pvrtc

Description

Name

VK_IMG_format_pvrtc - device extension

VK_IMG_format_pvrtc

Name String

VK_IMG_format_pvrtc

Extension Type

Device extension

Registered Extension Number

55

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Contact

Stuart Smith

Other Extension Metadata

Last Modified Date

2019-09-02

IP Status

Imagination Technologies Proprietary

Contributors

Stuart Smith, Imagination Technologies

Description

VK_IMG_format_pvrtc provides additional texture compression functionality specific to Imagination Technologies PowerVR Texture compression format (called PVRTC).

New Enum Constants

IMG_FORMAT_PVRTC_EXTENSION_NAME

IMG_FORMAT_PVRTC_SPEC_VERSION

Extending Format:

FORMAT_PVRTC1_2BPP_SRGB_BLOCK_IMG

FORMAT_PVRTC1_2BPP_UNORM_BLOCK_IMG

FORMAT_PVRTC1_4BPP_SRGB_BLOCK_IMG

FORMAT_PVRTC1_4BPP_UNORM_BLOCK_IMG

FORMAT_PVRTC2_2BPP_SRGB_BLOCK_IMG

FORMAT_PVRTC2_2BPP_UNORM_BLOCK_IMG

FORMAT_PVRTC2_4BPP_SRGB_BLOCK_IMG

FORMAT_PVRTC2_4BPP_UNORM_BLOCK_IMG

Version History

Revision 1, 2019-09-02 (Stuart Smith)

Initial version

See Also

No cross-references are available

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

type IMG_FORMAT_PVRTC_SPEC_VERSION = 1

pattern IMG_FORMAT_PVRTC_SPEC_VERSION :: forall a. Integral a => a

type IMG_FORMAT_PVRTC_EXTENSION_NAME = "VK_IMG_format_pvrtc"

pattern IMG_FORMAT_PVRTC_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a