Vulkan.Extensions.VK_INTEL_performance_query

Description

Name

VK_INTEL_performance_query - device extension

VK_INTEL_performance_query

Name String

VK_INTEL_performance_query

Extension Type

Device extension

Registered Extension Number

211

Revision

2

Extension and Version Dependencies

Requires Vulkan 1.0

Special Use

Developer tools

Contact

Lionel Landwerlin https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_INTEL_performance_query:%20&body=@llandwerlin%20

Other Extension Metadata

Last Modified Date

2018-05-16

IP Status

No known IP claims.

Contributors

Lionel Landwerlin, Intel

Piotr Maciejewski, Intel

Description

This extension allows an application to capture performance data to be interpreted by a external application or library.

Such a library is available at : https://github.com/intel/metrics-discovery

Performance analysis tools such as Graphics Performance Analyzers make use of this extension and the metrics-discovery library to present the data in a human readable way.

New Object Types

PerformanceConfigurationINTEL

New Commands

acquirePerformanceConfigurationINTEL

cmdSetPerformanceMarkerINTEL

cmdSetPerformanceOverrideINTEL

cmdSetPerformanceStreamMarkerINTEL

getPerformanceParameterINTEL

initializePerformanceApiINTEL

queueSetPerformanceConfigurationINTEL

releasePerformanceConfigurationINTEL

uninitializePerformanceApiINTEL

New Structures

InitializePerformanceApiInfoINTEL

PerformanceConfigurationAcquireInfoINTEL

PerformanceMarkerInfoINTEL

PerformanceOverrideInfoINTEL

PerformanceStreamMarkerInfoINTEL

PerformanceValueINTEL

Extending QueryPoolCreateInfo:

QueryPoolPerformanceQueryCreateInfoINTEL

New Unions

PerformanceValueDataINTEL

New Enums

PerformanceConfigurationTypeINTEL

PerformanceOverrideTypeINTEL

PerformanceParameterTypeINTEL

PerformanceValueTypeINTEL

QueryPoolSamplingModeINTEL

New Enum Constants

INTEL_PERFORMANCE_QUERY_EXTENSION_NAME

INTEL_PERFORMANCE_QUERY_SPEC_VERSION

Extending ObjectType:

OBJECT_TYPE_PERFORMANCE_CONFIGURATION_INTEL

Extending QueryType:

QUERY_TYPE_PERFORMANCE_QUERY_INTEL

Extending StructureType:

STRUCTURE_TYPE_INITIALIZE_PERFORMANCE_API_INFO_INTEL

STRUCTURE_TYPE_PERFORMANCE_CONFIGURATION_ACQUIRE_INFO_INTEL

STRUCTURE_TYPE_PERFORMANCE_MARKER_INFO_INTEL

STRUCTURE_TYPE_PERFORMANCE_OVERRIDE_INFO_INTEL

STRUCTURE_TYPE_PERFORMANCE_STREAM_MARKER_INFO_INTEL

STRUCTURE_TYPE_QUERY_POOL_PERFORMANCE_QUERY_CREATE_INFO_INTEL

Example Code

// A previously created device
VkDevice device;

// A queue derived from the device
VkQueue queue;

VkInitializePerformanceApiInfoINTEL performanceApiInfoIntel = {
  VK_STRUCTURE_TYPE_INITIALIZE_PERFORMANCE_API_INFO_INTEL,
  NULL,
  NULL
};

vkInitializePerformanceApiINTEL(
  device,
  &performanceApiInfoIntel);

VkQueryPoolPerformanceQueryCreateInfoINTEL queryPoolIntel = {
  VK_STRUCTURE_TYPE_QUERY_POOL_CREATE_INFO_INTEL,
  NULL,
  VK_QUERY_POOL_SAMPLING_MODE_MANUAL_INTEL,
};

VkQueryPoolCreateInfo queryPoolCreateInfo = {
  VK_STRUCTURE_TYPE_QUERY_POOL_CREATE_INFO,
  &queryPoolIntel,
  0,
  VK_QUERY_TYPE_PERFORMANCE_QUERY_INTEL,
  1,
  0
};

VkQueryPool queryPool;

VkResult result = vkCreateQueryPool(
  device,
  &queryPoolCreateInfo,
  NULL,
  &queryPool);

assert(VK_SUCCESS == result);

// A command buffer we want to record counters on
VkCommandBuffer commandBuffer;

VkCommandBufferBeginInfo commandBufferBeginInfo = {
  VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO,
  NULL,
  VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT,
  NULL
};

result = vkBeginCommandBuffer(commandBuffer, &commandBufferBeginInfo);

assert(VK_SUCCESS == result);

vkCmdResetQueryPool(
  commandBuffer,
  queryPool,
  0,
  1);

vkCmdBeginQuery(
  commandBuffer,
  queryPool,
  0,
  0);

// Perform the commands you want to get performance information on
// ...

// Perform a barrier to ensure all previous commands were complete before
// ending the query
vkCmdPipelineBarrier(commandBuffer,
  VK_PIPELINE_STAGE_BOTTOM_OF_PIPE_BIT,
  VK_PIPELINE_STAGE_BOTTOM_OF_PIPE_BIT,
  0,
  0,
  NULL,
  0,
  NULL,
  0,
  NULL);

vkCmdEndQuery(
  commandBuffer,
  queryPool,
  0);

result = vkEndCommandBuffer(commandBuffer);

assert(VK_SUCCESS == result);

VkPerformanceConfigurationAcquireInfoINTEL performanceConfigurationAcquireInfo = {
  VK_STRUCTURE_TYPE_PERFORMANCE_CONFIGURATION_ACQUIRE_INFO_INTEL,
  NULL,
  VK_PERFORMANCE_CONFIGURATION_TYPE_COMMAND_QUEUE_METRICS_DISCOVERY_ACTIVATED_INTEL
};

VkPerformanceConfigurationINTEL performanceConfigurationIntel;

result = vkAcquirePerformanceConfigurationINTEL(
  device,
  &performanceConfigurationAcquireInfo,
  &performanceConfigurationIntel);

vkQueueSetPerformanceConfigurationINTEL(queue, performanceConfigurationIntel);

assert(VK_SUCCESS == result);

// Submit the command buffer and wait for its completion
// ...

result = vkReleasePerformanceConfigurationINTEL(
  device,
  performanceConfigurationIntel);

assert(VK_SUCCESS == result);

// Get the report size from metrics-discovery's QueryReportSize

result = vkGetQueryPoolResults(
  device,
  queryPool,
  0, 1, QueryReportSize,
  data, QueryReportSize, 0);

assert(VK_SUCCESS == result);

// The data can then be passed back to metrics-discovery from which
// human readable values can be queried.

Version History

Revision 2, 2020-03-06 (Lionel Landwerlin)

Rename VkQueryPoolCreateInfoINTEL in VkQueryPoolPerformanceQueryCreateInfoINTEL

Revision 1, 2018-05-16 (Lionel Landwerlin)

Initial revision

See Also

InitializePerformanceApiInfoINTEL, PerformanceConfigurationAcquireInfoINTEL, PerformanceConfigurationINTEL, PerformanceConfigurationTypeINTEL, PerformanceMarkerInfoINTEL, PerformanceOverrideInfoINTEL, PerformanceOverrideTypeINTEL, PerformanceParameterTypeINTEL, PerformanceStreamMarkerInfoINTEL, PerformanceValueDataINTEL, PerformanceValueINTEL, PerformanceValueTypeINTEL, QueryPoolCreateInfoINTEL, QueryPoolPerformanceQueryCreateInfoINTEL, QueryPoolSamplingModeINTEL, acquirePerformanceConfigurationINTEL, cmdSetPerformanceMarkerINTEL, cmdSetPerformanceOverrideINTEL, cmdSetPerformanceStreamMarkerINTEL, getPerformanceParameterINTEL, initializePerformanceApiINTEL, queueSetPerformanceConfigurationINTEL, releasePerformanceConfigurationINTEL, uninitializePerformanceApiINTEL

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

initializePerformanceApiINTEL

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device used for the queries.

device must be a valid Device handle

-> ("initializeInfo" ::: InitializePerformanceApiInfoINTEL)
pInitializeInfo is a pointer to a InitializePerformanceApiInfoINTEL structure specifying initialization parameters.

pInitializeInfo must be a valid pointer to a valid InitializePerformanceApiInfoINTEL structure

-> io ()

vkInitializePerformanceApiINTEL - Initialize a device for performance queries

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

Device, InitializePerformanceApiInfoINTEL

uninitializePerformanceApiINTEL

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device used for the queries.

device must be a valid Device handle

-> io ()

vkUninitializePerformanceApiINTEL - Uninitialize a device for performance queries

Valid Usage (Implicit)

See Also

Device

cmdSetPerformanceMarkerINTEL :: forall io. MonadIO io => CommandBuffer -> PerformanceMarkerInfoINTEL -> io ()

vkCmdSetPerformanceMarkerINTEL - Markers

Parameters

The last marker set onto a command buffer before the end of a query will be part of the query result.

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pMarkerInfo must be a valid pointer to a valid PerformanceMarkerInfoINTEL structure

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, compute, or transfer operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute Transfer

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

CommandBuffer, PerformanceMarkerInfoINTEL

cmdSetPerformanceStreamMarkerINTEL :: forall io. MonadIO io => CommandBuffer -> PerformanceStreamMarkerInfoINTEL -> io ()

vkCmdSetPerformanceStreamMarkerINTEL - Markers

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pMarkerInfo must be a valid pointer to a valid PerformanceStreamMarkerInfoINTEL structure

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, compute, or transfer operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute Transfer

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

CommandBuffer, PerformanceStreamMarkerInfoINTEL

cmdSetPerformanceOverrideINTEL

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer where the override takes place.

-> PerformanceOverrideInfoINTEL
pOverrideInfo is a pointer to a PerformanceOverrideInfoINTEL structure selecting the parameter to override.

-> io ()

vkCmdSetPerformanceOverrideINTEL - Performance override settings

Valid Usage

pOverrideInfo must not be used with a PerformanceOverrideTypeINTEL that is not reported available by getPerformanceParameterINTEL

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pOverrideInfo must be a valid pointer to a valid PerformanceOverrideInfoINTEL structure

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, compute, or transfer operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute Transfer

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

CommandBuffer, PerformanceOverrideInfoINTEL

acquirePerformanceConfigurationINTEL

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that the performance query commands will be submitted to.

device must be a valid Device handle

-> PerformanceConfigurationAcquireInfoINTEL
pAcquireInfo is a pointer to a PerformanceConfigurationAcquireInfoINTEL structure, specifying the performance configuration to acquire.

pAcquireInfo must be a valid pointer to a valid PerformanceConfigurationAcquireInfoINTEL structure

-> io PerformanceConfigurationINTEL

vkAcquirePerformanceConfigurationINTEL - Acquire the performance query capability

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

Device, PerformanceConfigurationAcquireInfoINTEL, PerformanceConfigurationINTEL

releasePerformanceConfigurationINTEL

Arguments

:: forall io. MonadIO io

=> Device
device is the device associated to the configuration object to release.

-> PerformanceConfigurationINTEL
configuration is the configuration object to release.

-> io ()

vkReleasePerformanceConfigurationINTEL - Release a configuration to capture performance data

Valid Usage

configuration must not be released before all command buffers submitted while the configuration was set are in pending state

Valid Usage (Implicit)

device must be a valid Device handle

If configuration is not NULL_HANDLE, configuration must be a valid PerformanceConfigurationINTEL handle

If configuration is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to configuration must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

Device, PerformanceConfigurationINTEL

queueSetPerformanceConfigurationINTEL

Arguments

:: forall io. MonadIO io

=> Queue
queue is the queue on which the configuration will be used.

-> PerformanceConfigurationINTEL
configuration is the configuration to use.

-> io ()

vkQueueSetPerformanceConfigurationINTEL - Set a performance query

Valid Usage (Implicit)

queue must be a valid Queue handle

configuration must be a valid PerformanceConfigurationINTEL handle

Both of configuration, and queue must have been created, allocated, or retrieved from the same Device

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

- 	 - 	 Any 	 -

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

PerformanceConfigurationINTEL, Queue

getPerformanceParameterINTEL

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device to query.

device must be a valid Device handle

-> PerformanceParameterTypeINTEL
parameter is the parameter to query.

parameter must be a valid PerformanceParameterTypeINTEL value

-> io PerformanceValueINTEL

vkGetPerformanceParameterINTEL - Query performance capabilities of the device

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

Device, PerformanceParameterTypeINTEL, PerformanceValueINTEL

pattern STRUCTURE_TYPE_QUERY_POOL_CREATE_INFO_INTEL :: StructureType

data PerformanceValueINTEL

VkPerformanceValueINTEL - Container for value and types of parameters that can be queried

Valid Usage (Implicit)

type must be a valid PerformanceValueTypeINTEL value

If type is PERFORMANCE_VALUE_TYPE_STRING_INTEL, the valueString member of data must be a null-terminated UTF-8 string

See Also

PerformanceValueDataINTEL, PerformanceValueTypeINTEL, getPerformanceParameterINTEL

Constructors

PerformanceValueINTEL

Fields

type' :: PerformanceValueTypeINTEL

type is a PerformanceValueTypeINTEL value specifying the type of the returned data.

data' :: PerformanceValueDataINTEL

data is a PerformanceValueDataINTEL union specifying the value of the returned data.

data InitializePerformanceApiInfoINTEL

VkInitializePerformanceApiInfoINTEL - Structure specifying parameters of initialize of the device

Valid Usage (Implicit)

See Also

StructureType, initializePerformanceApiINTEL

Constructors

InitializePerformanceApiInfoINTEL

Fields

userData :: Ptr ()

pUserData is a pointer for application data.

data QueryPoolPerformanceQueryCreateInfoINTEL

VkQueryPoolPerformanceQueryCreateInfoINTEL - Structure specifying parameters to create a pool of performance queries

Members

To create a pool for Intel performance queries, set QueryPoolCreateInfo::queryType to QUERY_TYPE_PERFORMANCE_QUERY_INTEL and add a QueryPoolPerformanceQueryCreateInfoINTEL structure to the pNext chain of the QueryPoolCreateInfo structure.

Valid Usage (Implicit)

See Also

QueryPoolSamplingModeINTEL, StructureType

Constructors

QueryPoolPerformanceQueryCreateInfoINTEL

Fields

performanceCountersSampling :: QueryPoolSamplingModeINTEL

performanceCountersSampling describe how performance queries should be captured.

performanceCountersSampling must be a valid QueryPoolSamplingModeINTEL value

data PerformanceMarkerInfoINTEL

VkPerformanceMarkerInfoINTEL - Structure specifying performance markers

Valid Usage (Implicit)

See Also

StructureType, cmdSetPerformanceMarkerINTEL

Constructors

PerformanceMarkerInfoINTEL

Fields

marker :: Word64

marker is the marker value that will be recorded into the opaque query results.

data PerformanceStreamMarkerInfoINTEL

VkPerformanceStreamMarkerInfoINTEL - Structure specifying stream performance markers

Valid Usage

The value written by the application into marker must only used the valid bits as reported by getPerformanceParameterINTEL with the PERFORMANCE_PARAMETER_TYPE_STREAM_MARKER_VALID_BITS_INTEL

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PERFORMANCE_STREAM_MARKER_INFO_INTEL

pNext must be NULL

See Also

StructureType, cmdSetPerformanceStreamMarkerINTEL

Constructors

PerformanceStreamMarkerInfoINTEL

Fields

marker :: Word32

marker is the marker value that will be recorded into the reports consumed by an external application.

data PerformanceOverrideInfoINTEL

VkPerformanceOverrideInfoINTEL - Performance override info

Valid Usage (Implicit)

See Also

Bool32, PerformanceOverrideTypeINTEL, StructureType, cmdSetPerformanceOverrideINTEL

Constructors

PerformanceOverrideInfoINTEL

Fields

type' :: PerformanceOverrideTypeINTEL

type is the particular PerformanceOverrideTypeINTEL to set.

type must be a valid PerformanceOverrideTypeINTEL value

enable :: Bool

enable defines whether the override is enabled.

parameter :: Word64

parameter is a potential required parameter for the override.

data PerformanceConfigurationAcquireInfoINTEL

VkPerformanceConfigurationAcquireInfoINTEL - Acquire a configuration to capture performance data

Valid Usage (Implicit)

See Also

PerformanceConfigurationTypeINTEL, StructureType, acquirePerformanceConfigurationINTEL

Constructors

PerformanceConfigurationAcquireInfoINTEL

Fields

type' :: PerformanceConfigurationTypeINTEL

type is one of the PerformanceConfigurationTypeINTEL type of performance configuration that will be acquired.

type must be a valid PerformanceConfigurationTypeINTEL value

data PerformanceValueDataINTEL

Constructors

Value32 Word32

Value64 Word64

ValueFloat Float

ValueBool Bool

ValueString ByteString

peekPerformanceValueDataINTEL :: PerformanceValueTypeINTEL -> Ptr PerformanceValueDataINTEL -> IO PerformanceValueDataINTEL

newtype PerformanceConfigurationTypeINTEL

VkPerformanceConfigurationTypeINTEL - Type of performance configuration

See Also

PerformanceConfigurationAcquireInfoINTEL

Constructors

PerformanceConfigurationTypeINTEL Int32

Bundled Patterns

pattern PERFORMANCE_CONFIGURATION_TYPE_COMMAND_QUEUE_METRICS_DISCOVERY_ACTIVATED_INTEL :: PerformanceConfigurationTypeINTEL

newtype QueryPoolSamplingModeINTEL

VkQueryPoolSamplingModeINTEL - Enum specifying how performance queries should be captured

See Also

QueryPoolPerformanceQueryCreateInfoINTEL

Constructors

QueryPoolSamplingModeINTEL Int32

Bundled Patterns

pattern QUERY_POOL_SAMPLING_MODE_MANUAL_INTEL :: QueryPoolSamplingModeINTEL
QUERY_POOL_SAMPLING_MODE_MANUAL_INTEL is the default mode in which the application calls cmdBeginQuery and cmdEndQuery to record performance data.

newtype PerformanceOverrideTypeINTEL

VkPerformanceOverrideTypeINTEL - Performance override type

See Also

PerformanceOverrideInfoINTEL

Constructors

PerformanceOverrideTypeINTEL Int32

Bundled Patterns

pattern PERFORMANCE_OVERRIDE_TYPE_NULL_HARDWARE_INTEL :: PerformanceOverrideTypeINTEL
PERFORMANCE_OVERRIDE_TYPE_NULL_HARDWARE_INTEL turns all rendering operations into noop.

pattern PERFORMANCE_OVERRIDE_TYPE_FLUSH_GPU_CACHES_INTEL :: PerformanceOverrideTypeINTEL
PERFORMANCE_OVERRIDE_TYPE_FLUSH_GPU_CACHES_INTEL stalls the stream of commands until all previously emitted commands have completed and all caches been flushed and invalidated.

newtype PerformanceParameterTypeINTEL

VkPerformanceParameterTypeINTEL - Parameters that can be queried

See Also

getPerformanceParameterINTEL

Constructors

PerformanceParameterTypeINTEL Int32

Bundled Patterns

pattern PERFORMANCE_PARAMETER_TYPE_HW_COUNTERS_SUPPORTED_INTEL :: PerformanceParameterTypeINTEL
PERFORMANCE_PARAMETER_TYPE_HW_COUNTERS_SUPPORTED_INTEL has a boolean result which tells whether hardware counters can be captured.

pattern PERFORMANCE_PARAMETER_TYPE_STREAM_MARKER_VALID_BITS_INTEL :: PerformanceParameterTypeINTEL
PERFORMANCE_PARAMETER_TYPE_STREAM_MARKER_VALID_BITS_INTEL has a 32 bits integer result which tells how many bits can be written into the PerformanceValueINTEL value.

newtype PerformanceValueTypeINTEL

VkPerformanceValueTypeINTEL - Type of the parameters that can be queried

See Also

PerformanceValueINTEL

Constructors

PerformanceValueTypeINTEL Int32

Bundled Patterns

pattern PERFORMANCE_VALUE_TYPE_UINT32_INTEL :: PerformanceValueTypeINTEL

pattern PERFORMANCE_VALUE_TYPE_UINT64_INTEL :: PerformanceValueTypeINTEL

pattern PERFORMANCE_VALUE_TYPE_FLOAT_INTEL :: PerformanceValueTypeINTEL

pattern PERFORMANCE_VALUE_TYPE_BOOL_INTEL :: PerformanceValueTypeINTEL

pattern PERFORMANCE_VALUE_TYPE_STRING_INTEL :: PerformanceValueTypeINTEL

type QueryPoolCreateInfoINTEL = QueryPoolPerformanceQueryCreateInfoINTEL

type INTEL_PERFORMANCE_QUERY_SPEC_VERSION = 2

pattern INTEL_PERFORMANCE_QUERY_SPEC_VERSION :: forall a. Integral a => a

type INTEL_PERFORMANCE_QUERY_EXTENSION_NAME = "VK_INTEL_performance_query"

pattern INTEL_PERFORMANCE_QUERY_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype PerformanceConfigurationINTEL

VkPerformanceConfigurationINTEL - Device configuration for performance queries

See Also

acquirePerformanceConfigurationINTEL, queueSetPerformanceConfigurationINTEL, releasePerformanceConfigurationINTEL

Constructors

PerformanceConfigurationINTEL Word64