Vulkan.Extensions.VK_KHR_16bit_storage

Description

Name

VK_KHR_16bit_storage - device extension

VK_KHR_16bit_storage

Name String

VK_KHR_16bit_storage

Extension Type

Device extension

Registered Extension Number

84

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_get_physical_device_properties2

Requires VK_KHR_storage_buffer_storage_class

Deprecation state

Promoted to Vulkan 1.1

Contact

Jan-Harald Fredriksen https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_16bit_storage:%20&body=@janharaldfredriksen-arm%20

Other Extension Metadata

Last Modified Date

2017-09-05

IP Status

No known IP claims.

Interactions and External Dependencies

Promoted to Vulkan 1.1 Core

This extension requires SPV_KHR_16bit_storage

This extension provides API support for GL_EXT_shader_16bit_storage

Contributors

Alexander Galazin, ARM

Jan-Harald Fredriksen, ARM

Joerg Wagner, ARM

Neil Henning, Codeplay

Jeff Bolz, Nvidia

Daniel Koch, Nvidia

David Neto, Google

John Kessenich, Google

Description

The VK_KHR_16bit_storage extension allows use of 16-bit types in shader input and output interfaces, and push constant blocks. This extension introduces several new optional features which map to SPIR-V capabilities and allow access to 16-bit data in Block-decorated objects in the Uniform and the StorageBuffer storage classes, and objects in the PushConstant storage class. This extension allows 16-bit variables to be declared and used as user-defined shader inputs and outputs but does not change location assignment and component assignment rules.

Promotion to Vulkan 1.1

All functionality in this extension is included in core Vulkan 1.1, with the KHR suffix omitted. However, if Vulkan 1.1 is supported and this extension is not, the storageBuffer16BitAccess capability is optional. The original type, enum and command names are still available as aliases of the core functionality.

New Structures

Extending PhysicalDeviceFeatures2, DeviceCreateInfo:

PhysicalDevice16BitStorageFeaturesKHR

New Enum Constants

KHR_16BIT_STORAGE_EXTENSION_NAME

KHR_16BIT_STORAGE_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_PHYSICAL_DEVICE_16BIT_STORAGE_FEATURES_KHR

New SPIR-V Capabilities

StorageBuffer16BitAccess

UniformAndStorageBuffer16BitAccess

StoragePushConstant16

StorageInputOutput16

Version History

Revision 1, 2017-03-23 (Alexander Galazin)

Initial draft

See Also

PhysicalDevice16BitStorageFeaturesKHR

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_16BIT_STORAGE_FEATURES_KHR :: StructureType

type PhysicalDevice16BitStorageFeaturesKHR = PhysicalDevice16BitStorageFeatures

type KHR_16BIT_STORAGE_SPEC_VERSION = 1

pattern KHR_16BIT_STORAGE_SPEC_VERSION :: forall a. Integral a => a

type KHR_16BIT_STORAGE_EXTENSION_NAME = "VK_KHR_16bit_storage"

pattern KHR_16BIT_STORAGE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a