Vulkan.Extensions.VK_KHR_bind_memory2

Description

Name

VK_KHR_bind_memory2 - device extension

VK_KHR_bind_memory2

Name String

VK_KHR_bind_memory2

Extension Type

Device extension

Registered Extension Number

158

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Deprecation state

Promoted to Vulkan 1.1

Contact

Tobias Hector https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_bind_memory2:%20&body=@tobski%20

Other Extension Metadata

Last Modified Date

2017-09-05

IP Status

No known IP claims.

Interactions and External Dependencies

Promoted to Vulkan 1.1 Core

Contributors

Jeff Bolz, NVIDIA

Tobias Hector, Imagination Technologies

Description

This extension provides versions of bindBufferMemory and bindImageMemory that allow multiple bindings to be performed at once, and are extensible.

This extension also introduces IMAGE_CREATE_ALIAS_BIT_KHR, which allows “identical” images that alias the same memory to interpret the contents consistently, even across image layout changes.

Promotion to Vulkan 1.1

All functionality in this extension is included in core Vulkan 1.1, with the KHR suffix omitted. The original type, enum and command names are still available as aliases of the core functionality.

New Commands

bindBufferMemory2KHR

bindImageMemory2KHR

New Structures

BindBufferMemoryInfoKHR

BindImageMemoryInfoKHR

New Enum Constants

KHR_BIND_MEMORY_2_EXTENSION_NAME

KHR_BIND_MEMORY_2_SPEC_VERSION

Extending ImageCreateFlagBits:

IMAGE_CREATE_ALIAS_BIT_KHR

Extending StructureType:

STRUCTURE_TYPE_BIND_BUFFER_MEMORY_INFO_KHR

STRUCTURE_TYPE_BIND_IMAGE_MEMORY_INFO_KHR

Version History

Revision 1, 2017-05-19 (Tobias Hector)

Pulled bind memory functions into their own extension

See Also

BindBufferMemoryInfoKHR, BindImageMemoryInfoKHR, bindBufferMemory2KHR, bindImageMemory2KHR

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

pattern STRUCTURE_TYPE_BIND_BUFFER_MEMORY_INFO_KHR :: StructureType

pattern STRUCTURE_TYPE_BIND_IMAGE_MEMORY_INFO_KHR :: StructureType

pattern IMAGE_CREATE_ALIAS_BIT_KHR :: ImageCreateFlagBits

bindBufferMemory2KHR :: MonadIO io => Device -> ("bindInfos" ::: Vector (SomeStruct BindBufferMemoryInfo)) -> io ()

bindImageMemory2KHR :: MonadIO io => Device -> ("bindInfos" ::: Vector (SomeStruct BindImageMemoryInfo)) -> io ()

type BindBufferMemoryInfoKHR = BindBufferMemoryInfo

type BindImageMemoryInfoKHR = BindImageMemoryInfo

type KHR_BIND_MEMORY_2_SPEC_VERSION = 1

pattern KHR_BIND_MEMORY_2_SPEC_VERSION :: forall a. Integral a => a

type KHR_BIND_MEMORY_2_EXTENSION_NAME = "VK_KHR_bind_memory2"

pattern KHR_BIND_MEMORY_2_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a