Vulkan.Extensions.VK_KHR_depth_stencil_resolve

Description

Name

VK_KHR_depth_stencil_resolve - device extension

VK_KHR_depth_stencil_resolve

Name String

VK_KHR_depth_stencil_resolve

Extension Type

Device extension

Registered Extension Number

200

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_create_renderpass2

Deprecation state

Promoted to Vulkan 1.2

Contact

Jan-Harald Fredriksen https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_depth_stencil_resolve:%20&body=@janharald%20

Other Extension Metadata

Last Modified Date

2018-04-09

Interactions and External Dependencies

Promoted to Vulkan 1.2 Core

Contributors

Jan-Harald Fredriksen, Arm

Andrew Garrard, Samsung Electronics

Soowan Park, Samsung Electronics

Jeff Bolz, NVIDIA

Daniel Rakos, AMD

Description

This extension adds support for automatically resolving multisampled depth/stencil attachments in a subpass in a similar manner as for color attachments.

Multisampled color attachments can be resolved at the end of a subpass by specifying pResolveAttachments entries corresponding to the pColorAttachments array entries. This does not allow for a way to map the resolve attachments to the depth/stencil attachment. The cmdResolveImage command does not allow for depth/stencil images. While there are other ways to resolve the depth/stencil attachment, they can give sub-optimal performance. Extending the SubpassDescription2 in this extension allows an application to add a pDepthStencilResolveAttachment, that is similar to the color pResolveAttachments, that the pDepthStencilAttachment can be resolved into.

Depth and stencil samples are resolved to a single value based on the resolve mode. The set of possible resolve modes is defined in the ResolveModeFlagBits enum. The RESOLVE_MODE_SAMPLE_ZERO_BIT mode is the only mode that is required of all implementations (that support the extension or support Vulkan 1.2 or higher). Some implementations may also support averaging (the same as color sample resolve) or taking the minimum or maximum sample, which may be more suitable for depth/stencil resolve.

Promotion to Vulkan 1.2

All functionality in this extension is included in core Vulkan 1.2, with the KHR suffix omitted. The original type, enum and command names are still available as aliases of the core functionality.

New Structures

Extending PhysicalDeviceProperties2:

PhysicalDeviceDepthStencilResolvePropertiesKHR

Extending SubpassDescription2:

SubpassDescriptionDepthStencilResolveKHR

New Enums

ResolveModeFlagBitsKHR

New Bitmasks

ResolveModeFlagsKHR

New Enum Constants

KHR_DEPTH_STENCIL_RESOLVE_EXTENSION_NAME

KHR_DEPTH_STENCIL_RESOLVE_SPEC_VERSION

Extending ResolveModeFlagBits:

RESOLVE_MODE_AVERAGE_BIT_KHR

RESOLVE_MODE_MAX_BIT_KHR

RESOLVE_MODE_MIN_BIT_KHR

RESOLVE_MODE_NONE_KHR

RESOLVE_MODE_SAMPLE_ZERO_BIT_KHR

Extending StructureType:

STRUCTURE_TYPE_PHYSICAL_DEVICE_DEPTH_STENCIL_RESOLVE_PROPERTIES_KHR

STRUCTURE_TYPE_SUBPASS_DESCRIPTION_DEPTH_STENCIL_RESOLVE_KHR

Version History

Revision 1, 2018-04-09 (Jan-Harald Fredriksen)

Initial revision

See Also

PhysicalDeviceDepthStencilResolvePropertiesKHR, ResolveModeFlagBitsKHR, ResolveModeFlagsKHR, SubpassDescriptionDepthStencilResolveKHR

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_DEPTH_STENCIL_RESOLVE_PROPERTIES_KHR :: StructureType

pattern STRUCTURE_TYPE_SUBPASS_DESCRIPTION_DEPTH_STENCIL_RESOLVE_KHR :: StructureType

pattern RESOLVE_MODE_NONE_KHR :: ResolveModeFlagBits

pattern RESOLVE_MODE_SAMPLE_ZERO_BIT_KHR :: ResolveModeFlagBits

pattern RESOLVE_MODE_AVERAGE_BIT_KHR :: ResolveModeFlagBits

pattern RESOLVE_MODE_MIN_BIT_KHR :: ResolveModeFlagBits

pattern RESOLVE_MODE_MAX_BIT_KHR :: ResolveModeFlagBits

type ResolveModeFlagsKHR = ResolveModeFlags

type ResolveModeFlagBitsKHR = ResolveModeFlagBits

type PhysicalDeviceDepthStencilResolvePropertiesKHR = PhysicalDeviceDepthStencilResolveProperties

type SubpassDescriptionDepthStencilResolveKHR = SubpassDescriptionDepthStencilResolve

type KHR_DEPTH_STENCIL_RESOLVE_SPEC_VERSION = 1

pattern KHR_DEPTH_STENCIL_RESOLVE_SPEC_VERSION :: forall a. Integral a => a

type KHR_DEPTH_STENCIL_RESOLVE_EXTENSION_NAME = "VK_KHR_depth_stencil_resolve"

pattern KHR_DEPTH_STENCIL_RESOLVE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a