Vulkan.Extensions.VK_KHR_descriptor_update_template

Documentation

pattern STRUCTURE_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_CREATE_INFO_KHR :: StructureType

pattern OBJECT_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_KHR :: ObjectType

pattern DESCRIPTOR_UPDATE_TEMPLATE_TYPE_DESCRIPTOR_SET_KHR :: DescriptorUpdateTemplateType

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_KHR_EXT :: DebugReportObjectTypeEXT

createDescriptorUpdateTemplateKHR :: MonadIO io => Device -> DescriptorUpdateTemplateCreateInfo -> ("allocator" ::: Maybe AllocationCallbacks) -> io DescriptorUpdateTemplate

destroyDescriptorUpdateTemplateKHR :: MonadIO io => Device -> DescriptorUpdateTemplate -> ("allocator" ::: Maybe AllocationCallbacks) -> io ()

updateDescriptorSetWithTemplateKHR :: MonadIO io => Device -> DescriptorSet -> DescriptorUpdateTemplate -> ("data" ::: Ptr ()) -> io ()

type DescriptorUpdateTemplateCreateFlagsKHR = DescriptorUpdateTemplateCreateFlags

type DescriptorUpdateTemplateKHR = DescriptorUpdateTemplate

type DescriptorUpdateTemplateTypeKHR = DescriptorUpdateTemplateType

type DescriptorUpdateTemplateEntryKHR = DescriptorUpdateTemplateEntry

type DescriptorUpdateTemplateCreateInfoKHR = DescriptorUpdateTemplateCreateInfo

type KHR_DESCRIPTOR_UPDATE_TEMPLATE_SPEC_VERSION = 1

pattern KHR_DESCRIPTOR_UPDATE_TEMPLATE_SPEC_VERSION :: forall a. Integral a => a

type KHR_DESCRIPTOR_UPDATE_TEMPLATE_EXTENSION_NAME = "VK_KHR_descriptor_update_template"

pattern KHR_DESCRIPTOR_UPDATE_TEMPLATE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

cmdPushDescriptorSetWithTemplateKHR

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer that the descriptors will be recorded in.

-> DescriptorUpdateTemplate
descriptorUpdateTemplate is a descriptor update template defining how to interpret the descriptor information in pData.

-> PipelineLayout
layout is a PipelineLayout object used to program the bindings. It must be compatible with the layout used to create the descriptorUpdateTemplate handle.

-> ("set" ::: Word32)
set is the set number of the descriptor set in the pipeline layout that will be updated. This must be the same number used to create the descriptorUpdateTemplate handle.

-> ("data" ::: Ptr ())
pData is a pointer to memory containing descriptors for the templated update.

-> io ()

vkCmdPushDescriptorSetWithTemplateKHR - Pushes descriptor updates into a command buffer using a descriptor update template

Valid Usage

The pipelineBindPoint specified during the creation of the descriptor update template must be supported by the commandBuffer’s parent CommandPool’s queue family

pData must be a valid pointer to a memory containing one or more valid instances of DescriptorImageInfo, DescriptorBufferInfo, or BufferView in a layout defined by descriptorUpdateTemplate when it was created with createDescriptorUpdateTemplateKHR

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

descriptorUpdateTemplate must be a valid DescriptorUpdateTemplate handle

layout must be a valid PipelineLayout handle

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, or compute operations

Each of commandBuffer, descriptorUpdateTemplate, and layout must have been created, allocated, or retrieved from the same Device

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute

API example.

struct AppDataStructure
{
    VkDescriptorImageInfo  imageInfo;          // a single image info
    // ... some more application related data
};

const VkDescriptorUpdateTemplateEntry descriptorUpdateTemplateEntries[] =
{
    // binding to a single image descriptor
    {
        0,                                           // binding
        0,                                           // dstArrayElement
        1,                                           // descriptorCount
        VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER,   // descriptorType
        offsetof(AppDataStructure, imageInfo),       // offset
        0                                            // stride is not required if descriptorCount is 1
    }
};

// create a descriptor update template for descriptor set updates
const VkDescriptorUpdateTemplateCreateInfo createInfo =
{
    VK_STRUCTURE_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_CREATE_INFO,  // sType
    NULL,                                                      // pNext
    0,                                                         // flags
    1,                                                         // descriptorUpdateEntryCount
    descriptorUpdateTemplateEntries,                           // pDescriptorUpdateEntries
    VK_DESCRIPTOR_UPDATE_TEMPLATE_TYPE_PUSH_DESCRIPTORS_KHR,   // templateType
    0,                                                         // descriptorSetLayout, ignored by given templateType
    VK_PIPELINE_BIND_POINT_GRAPHICS,                           // pipelineBindPoint
    myPipelineLayout,                                          // pipelineLayout
    0,                                                         // set
};

VkDescriptorUpdateTemplate myDescriptorUpdateTemplate;
myResult = vkCreateDescriptorUpdateTemplate(
    myDevice,
    &createInfo,
    NULL,
    &myDescriptorUpdateTemplate);
}

AppDataStructure appData;
// fill appData here or cache it in your engine
vkCmdPushDescriptorSetWithTemplateKHR(myCmdBuffer, myDescriptorUpdateTemplate, myPipelineLayout, 0,&appData);

See Also

CommandBuffer, DescriptorUpdateTemplate, PipelineLayout

newtype DebugReportObjectTypeEXT

VkDebugReportObjectTypeEXT - Specify the type of an object handle

Description

'

DebugReportObjectTypeEXT 	 Vulkan Handle Type

DEBUG_REPORT_OBJECT_TYPE_UNKNOWN_EXT 	 Unknown/Undefined Handle

DEBUG_REPORT_OBJECT_TYPE_INSTANCE_EXT 	 Instance

DEBUG_REPORT_OBJECT_TYPE_PHYSICAL_DEVICE_EXT 	 PhysicalDevice

DEBUG_REPORT_OBJECT_TYPE_DEVICE_EXT 	 Device

DEBUG_REPORT_OBJECT_TYPE_QUEUE_EXT 	 Queue

DEBUG_REPORT_OBJECT_TYPE_SEMAPHORE_EXT 	 Semaphore

DEBUG_REPORT_OBJECT_TYPE_COMMAND_BUFFER_EXT 	 CommandBuffer

DEBUG_REPORT_OBJECT_TYPE_FENCE_EXT 	 Fence

DEBUG_REPORT_OBJECT_TYPE_DEVICE_MEMORY_EXT 	 DeviceMemory

DEBUG_REPORT_OBJECT_TYPE_BUFFER_EXT 	 Buffer

DEBUG_REPORT_OBJECT_TYPE_IMAGE_EXT 	 Image

DEBUG_REPORT_OBJECT_TYPE_EVENT_EXT 	 Event

DEBUG_REPORT_OBJECT_TYPE_QUERY_POOL_EXT 	 QueryPool

DEBUG_REPORT_OBJECT_TYPE_BUFFER_VIEW_EXT 	 BufferView

DEBUG_REPORT_OBJECT_TYPE_IMAGE_VIEW_EXT 	 ImageView

DEBUG_REPORT_OBJECT_TYPE_SHADER_MODULE_EXT 	 ShaderModule

DEBUG_REPORT_OBJECT_TYPE_PIPELINE_CACHE_EXT 	 PipelineCache

DEBUG_REPORT_OBJECT_TYPE_PIPELINE_LAYOUT_EXT 	 PipelineLayout

DEBUG_REPORT_OBJECT_TYPE_RENDER_PASS_EXT 	 RenderPass

DEBUG_REPORT_OBJECT_TYPE_PIPELINE_EXT 	 Pipeline

DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_LAYOUT_EXT 	 DescriptorSetLayout

DEBUG_REPORT_OBJECT_TYPE_SAMPLER_EXT 	 Sampler

DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_POOL_EXT 	 DescriptorPool

DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_EXT 	 DescriptorSet

DEBUG_REPORT_OBJECT_TYPE_FRAMEBUFFER_EXT 	 Framebuffer

DEBUG_REPORT_OBJECT_TYPE_COMMAND_POOL_EXT 	 CommandPool

DEBUG_REPORT_OBJECT_TYPE_SURFACE_KHR_EXT 	 SurfaceKHR

DEBUG_REPORT_OBJECT_TYPE_SWAPCHAIN_KHR_EXT 	 SwapchainKHR

DEBUG_REPORT_OBJECT_TYPE_DEBUG_REPORT_CALLBACK_EXT_EXT 	 DebugReportCallbackEXT

DEBUG_REPORT_OBJECT_TYPE_DISPLAY_KHR_EXT 	 DisplayKHR

DEBUG_REPORT_OBJECT_TYPE_DISPLAY_MODE_KHR_EXT 	 DisplayModeKHR

DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_EXT 	 DescriptorUpdateTemplate

DebugReportObjectTypeEXT and Vulkan Handle Relationship

Note

The primary expected use of ERROR_VALIDATION_FAILED_EXT is for validation layer testing. It is not expected that an application would see this error code during normal use of the validation layers.

See Also

DebugMarkerObjectNameInfoEXT, DebugMarkerObjectTagInfoEXT, debugReportMessageEXT

Constructors

DebugReportObjectTypeEXT Int32

Bundled Patterns

pattern DEBUG_REPORT_OBJECT_TYPE_UNKNOWN_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_INSTANCE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_PHYSICAL_DEVICE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DEVICE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_QUEUE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SEMAPHORE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_COMMAND_BUFFER_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_FENCE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DEVICE_MEMORY_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_BUFFER_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_IMAGE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_EVENT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_QUERY_POOL_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_BUFFER_VIEW_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_IMAGE_VIEW_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SHADER_MODULE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_PIPELINE_CACHE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_PIPELINE_LAYOUT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_RENDER_PASS_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_PIPELINE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_LAYOUT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SAMPLER_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_POOL_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_FRAMEBUFFER_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_COMMAND_POOL_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SURFACE_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SWAPCHAIN_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DEBUG_REPORT_CALLBACK_EXT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DISPLAY_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DISPLAY_MODE_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_VALIDATION_CACHE_EXT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_ACCELERATION_STRUCTURE_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_EXT :: DebugReportObjectTypeEXT