Vulkan.Extensions.VK_KHR_draw_indirect_count

Description

Name

VK_KHR_draw_indirect_count - device extension

VK_KHR_draw_indirect_count

Name String

VK_KHR_draw_indirect_count

Extension Type

Device extension

Registered Extension Number

170

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Deprecation state

Promoted to Vulkan 1.2

Contact

Piers Daniell https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_draw_indirect_count:%20&body=@pdaniell-nv%20

Other Extension Metadata

Last Modified Date

2017-08-25

Interactions and External Dependencies

Promoted to Vulkan 1.2 Core

IP Status

No known IP claims.

Contributors

Matthaeus G. Chajdas, AMD

Derrick Owens, AMD

Graham Sellers, AMD

Daniel Rakos, AMD

Dominik Witczak, AMD

Piers Daniell, NVIDIA

Description

This extension is based off the VK_AMD_draw_indirect_count extension. This extension allows an application to source the number of draw calls for indirect draw calls from a buffer.

Applications might want to do culling on the GPU via a compute shader prior to the draw. This enables the application to generate arbitrary amounts of draw commands and execute them without host intervention.

Promotion to Vulkan 1.2

All functionality in this extension is included in core Vulkan 1.2, with the KHR suffix omitted. However, if Vulkan 1.2 is supported and this extension is not, the entry points cmdDrawIndirectCount and cmdDrawIndexedIndirectCount are optional. The original type, enum and command names are still available as aliases of the core functionality.

New Commands

cmdDrawIndexedIndirectCountKHR

cmdDrawIndirectCountKHR

New Enum Constants

KHR_DRAW_INDIRECT_COUNT_EXTENSION_NAME

KHR_DRAW_INDIRECT_COUNT_SPEC_VERSION

Version History

Revision 1, 2017-08-25 (Piers Daniell)

Initial draft based off VK_AMD_draw_indirect_count

See Also

cmdDrawIndexedIndirectCountKHR, cmdDrawIndirectCountKHR

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

cmdDrawIndirectCountKHR :: MonadIO io => CommandBuffer -> Buffer -> ("offset" ::: DeviceSize) -> ("countBuffer" ::: Buffer) -> ("countBufferOffset" ::: DeviceSize) -> ("maxDrawCount" ::: Word32) -> ("stride" ::: Word32) -> io ()

cmdDrawIndexedIndirectCountKHR :: MonadIO io => CommandBuffer -> Buffer -> ("offset" ::: DeviceSize) -> ("countBuffer" ::: Buffer) -> ("countBufferOffset" ::: DeviceSize) -> ("maxDrawCount" ::: Word32) -> ("stride" ::: Word32) -> io ()

type KHR_DRAW_INDIRECT_COUNT_SPEC_VERSION = 1

pattern KHR_DRAW_INDIRECT_COUNT_SPEC_VERSION :: forall a. Integral a => a

type KHR_DRAW_INDIRECT_COUNT_EXTENSION_NAME = "VK_KHR_draw_indirect_count"

pattern KHR_DRAW_INDIRECT_COUNT_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a