Vulkan.Extensions.VK_KHR_driver_properties

Description

Name

VK_KHR_driver_properties - device extension

VK_KHR_driver_properties

Name String

VK_KHR_driver_properties

Extension Type

Device extension

Registered Extension Number

197

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_get_physical_device_properties2

Deprecation state

Promoted to Vulkan 1.2

Contact

Daniel Rakos https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_driver_properties:%20&body=@drakos-amd%20

Other Extension Metadata

Last Modified Date

2018-04-11

Interactions and External Dependencies

Promoted to Vulkan 1.2 Core

IP Status

No known IP claims.

Contributors

Baldur Karlsson

Matthaeus G. Chajdas, AMD

Piers Daniell, NVIDIA

Alexander Galazin, Arm

Jesse Hall, Google

Daniel Rakos, AMD

Description

This extension provides a new physical device query which allows retrieving information about the driver implementation, allowing applications to determine which physical device corresponds to which particular vendor’s driver, and which conformance test suite version the driver implementation is compliant with.

Promotion to Vulkan 1.2

All functionality in this extension is included in core Vulkan 1.2, with the KHR suffix omitted. The original type, enum and command names are still available as aliases of the core functionality.

New Structures

ConformanceVersionKHR

Extending PhysicalDeviceProperties2:

PhysicalDeviceDriverPropertiesKHR

New Enums

DriverIdKHR

New Enum Constants

KHR_DRIVER_PROPERTIES_EXTENSION_NAME

KHR_DRIVER_PROPERTIES_SPEC_VERSION

MAX_DRIVER_INFO_SIZE_KHR

MAX_DRIVER_NAME_SIZE_KHR

Extending DriverId:

DRIVER_ID_AMD_OPEN_SOURCE_KHR

DRIVER_ID_AMD_PROPRIETARY_KHR

DRIVER_ID_ARM_PROPRIETARY_KHR

DRIVER_ID_BROADCOM_PROPRIETARY_KHR

DRIVER_ID_GGP_PROPRIETARY_KHR

DRIVER_ID_GOOGLE_SWIFTSHADER_KHR

DRIVER_ID_IMAGINATION_PROPRIETARY_KHR

DRIVER_ID_INTEL_OPEN_SOURCE_MESA_KHR

DRIVER_ID_INTEL_PROPRIETARY_WINDOWS_KHR

DRIVER_ID_MESA_RADV_KHR

DRIVER_ID_NVIDIA_PROPRIETARY_KHR

DRIVER_ID_QUALCOMM_PROPRIETARY_KHR

Extending StructureType:

STRUCTURE_TYPE_PHYSICAL_DEVICE_DRIVER_PROPERTIES_KHR

Version History

Revision 1, 2018-04-11 (Daniel Rakos)

Internal revisions

See Also

MAX_DRIVER_INFO_SIZE_KHR, MAX_DRIVER_NAME_SIZE_KHR, ConformanceVersionKHR, DriverIdKHR, PhysicalDeviceDriverPropertiesKHR

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_DRIVER_PROPERTIES_KHR :: StructureType

pattern DRIVER_ID_AMD_PROPRIETARY_KHR :: DriverId

pattern DRIVER_ID_AMD_OPEN_SOURCE_KHR :: DriverId

pattern DRIVER_ID_MESA_RADV_KHR :: DriverId

pattern DRIVER_ID_NVIDIA_PROPRIETARY_KHR :: DriverId

pattern DRIVER_ID_INTEL_PROPRIETARY_WINDOWS_KHR :: DriverId

pattern DRIVER_ID_INTEL_OPEN_SOURCE_MESA_KHR :: DriverId

pattern DRIVER_ID_IMAGINATION_PROPRIETARY_KHR :: DriverId

pattern DRIVER_ID_QUALCOMM_PROPRIETARY_KHR :: DriverId

pattern DRIVER_ID_ARM_PROPRIETARY_KHR :: DriverId

pattern DRIVER_ID_GOOGLE_SWIFTSHADER_KHR :: DriverId

pattern DRIVER_ID_GGP_PROPRIETARY_KHR :: DriverId

pattern DRIVER_ID_BROADCOM_PROPRIETARY_KHR :: DriverId

pattern MAX_DRIVER_NAME_SIZE_KHR :: Integral a => a

pattern MAX_DRIVER_INFO_SIZE_KHR :: Integral a => a

type DriverIdKHR = DriverId

type ConformanceVersionKHR = ConformanceVersion

type PhysicalDeviceDriverPropertiesKHR = PhysicalDeviceDriverProperties

type KHR_DRIVER_PROPERTIES_SPEC_VERSION = 1

pattern KHR_DRIVER_PROPERTIES_SPEC_VERSION :: forall a. Integral a => a

type KHR_DRIVER_PROPERTIES_EXTENSION_NAME = "VK_KHR_driver_properties"

pattern KHR_DRIVER_PROPERTIES_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a