Vulkan.Extensions.VK_KHR_external_fence

Description

Name

VK_KHR_external_fence - device extension

VK_KHR_external_fence

Name String

VK_KHR_external_fence

Extension Type

Device extension

Registered Extension Number

114

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_external_fence_capabilities

Deprecation state

Promoted to Vulkan 1.1

Contact

Jesse Hall https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_external_fence:%20&body=@critsec%20

Other Extension Metadata

Last Modified Date

2017-05-08

IP Status

No known IP claims.

Interactions and External Dependencies

Promoted to Vulkan 1.1 Core

Contributors

Jesse Hall, Google

James Jones, NVIDIA

Jeff Juliano, NVIDIA

Cass Everitt, Oculus

Contributors to VK_KHR_external_semaphore

Description

An application using external memory may wish to synchronize access to that memory using fences. This extension enables an application to create fences from which non-Vulkan handles that reference the underlying synchronization primitive can be exported.

Promotion to Vulkan 1.1

All functionality in this extension is included in core Vulkan 1.1, with the KHR suffix omitted. The original type, enum and command names are still available as aliases of the core functionality.

New Structures

Extending FenceCreateInfo:

ExportFenceCreateInfoKHR

New Enums

FenceImportFlagBitsKHR

New Bitmasks

FenceImportFlagsKHR

New Enum Constants

KHR_EXTERNAL_FENCE_EXTENSION_NAME

KHR_EXTERNAL_FENCE_SPEC_VERSION

Extending FenceImportFlagBits:

FENCE_IMPORT_TEMPORARY_BIT_KHR

Extending StructureType:

STRUCTURE_TYPE_EXPORT_FENCE_CREATE_INFO_KHR

Issues

This extension borrows concepts, semantics, and language from VK_KHR_external_semaphore. That extension’s issues apply equally to this extension.

Version History

Revision 1, 2017-05-08 (Jesse Hall)

Initial revision

See Also

ExportFenceCreateInfoKHR, FenceImportFlagBitsKHR, FenceImportFlagsKHR

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

pattern STRUCTURE_TYPE_EXPORT_FENCE_CREATE_INFO_KHR :: StructureType

pattern FENCE_IMPORT_TEMPORARY_BIT_KHR :: FenceImportFlagBits

type FenceImportFlagsKHR = FenceImportFlags

type FenceImportFlagBitsKHR = FenceImportFlagBits

type ExportFenceCreateInfoKHR = ExportFenceCreateInfo

type KHR_EXTERNAL_FENCE_SPEC_VERSION = 1

pattern KHR_EXTERNAL_FENCE_SPEC_VERSION :: forall a. Integral a => a

type KHR_EXTERNAL_FENCE_EXTENSION_NAME = "VK_KHR_external_fence"

pattern KHR_EXTERNAL_FENCE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a