Vulkan.Extensions.VK_KHR_external_fence_win32

Documentation

getFenceWin32HandleKHR

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that created the fence being exported.

device must be a valid Device handle

-> FenceGetWin32HandleInfoKHR
pGetWin32HandleInfo is a pointer to a FenceGetWin32HandleInfoKHR structure containing parameters of the export operation.

pGetWin32HandleInfo must be a valid pointer to a valid FenceGetWin32HandleInfoKHR structure

-> io HANDLE

vkGetFenceWin32HandleKHR - Get a Windows HANDLE for a fence

Description

For handle types defined as NT handles, the handles returned by getFenceWin32HandleKHR are owned by the application. To avoid leaking resources, the application must release ownership of them using the CloseHandle system call when they are no longer needed.

Exporting a Windows handle from a fence may have side effects depending on the transference of the specified handle type, as described in Importing Fence Payloads.

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

Device, FenceGetWin32HandleInfoKHR

importFenceWin32HandleKHR

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that created the fence.

device must be a valid Device handle

-> ImportFenceWin32HandleInfoKHR
pImportFenceWin32HandleInfo is a pointer to a ImportFenceWin32HandleInfoKHR structure specifying the fence and import parameters.

pImportFenceWin32HandleInfo must be a valid pointer to a valid ImportFenceWin32HandleInfoKHR structure

-> io ()

vkImportFenceWin32HandleKHR - Import a fence from a Windows HANDLE

Description

Importing a fence payload from Windows handles does not transfer ownership of the handle to the Vulkan implementation. For handle types defined as NT handles, the application must release ownership using the CloseHandle system call when the handle is no longer needed.

Applications can import the same fence payload into multiple instances of Vulkan, into the same instance from which it was exported, and multiple times into a given Vulkan instance.

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_INVALID_EXTERNAL_HANDLE

See Also

Device, ImportFenceWin32HandleInfoKHR

data ImportFenceWin32HandleInfoKHR

VkImportFenceWin32HandleInfoKHR - (None)

Description

The handle types supported by handleType are:

Handle Type 	 Transference 	 Permanence Supported

EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_BIT 	 Reference 	 Temporary,Permanent

EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT 	 Reference 	 Temporary,Permanent

Handle Types Supported by ImportFenceWin32HandleInfoKHR

Valid Usage

handleType must be a value included in the Handle Types Supported by table

If handleType is not EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_BIT, name must be NULL

If handleType is not 0 and handle is NULL, name must name a valid synchronization primitive of the type specified by handleType

If handleType is not 0 and name is NULL, handle must be a valid handle of the type specified by handleType

If handle is not NULL, name must be NULL

If handle is not NULL, it must obey any requirements listed for handleType in external fence handle types compatibility

If name is not NULL, it must obey any requirements listed for handleType in external fence handle types compatibility

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_IMPORT_FENCE_WIN32_HANDLE_INFO_KHR

pNext must be NULL

fence must be a valid Fence handle

flags must be a valid combination of FenceImportFlagBits values

If handleType is not 0, handleType must be a valid ExternalFenceHandleTypeFlagBits value

Host Synchronization

Host access to fence must be externally synchronized

See Also

ExternalFenceHandleTypeFlagBits, Fence, FenceImportFlags, StructureType, importFenceWin32HandleKHR

Constructors

ImportFenceWin32HandleInfoKHR

Fields

fence :: Fence

fence is the fence into which the state will be imported.

flags :: FenceImportFlags

flags is a bitmask of FenceImportFlagBits specifying additional parameters for the fence payload import operation.

handleType :: ExternalFenceHandleTypeFlagBits

handleType specifies the type of handle.

handle :: HANDLE

handle is the external handle to import, or NULL.

name :: LPCWSTR

name is a null-terminated UTF-16 string naming the underlying synchronization primitive to import, or NULL.

data ExportFenceWin32HandleInfoKHR

VkExportFenceWin32HandleInfoKHR - Structure specifying additional attributes of Windows handles exported from a fence

Description

If ExportFenceCreateInfo is not present in the same pNext chain, this structure is ignored.

If ExportFenceCreateInfo is present in the pNext chain of FenceCreateInfo with a Windows handleType, but either ExportFenceWin32HandleInfoKHR is not present in the pNext chain, or if it is but pAttributes is set to NULL, default security descriptor values will be used, and child processes created by the application will not inherit the handle, as described in the MSDN documentation for “Synchronization Object Security and Access Rights”1. Further, if the structure is not present, the access rights will be

DXGI_SHARED_RESOURCE_READ | DXGI_SHARED_RESOURCE_WRITE

for handles of the following types:

EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_BIT

1

https://docs.microsoft.com/en-us/windows/win32/sync/synchronization-object-security-and-access-rights

Valid Usage

If ExportFenceCreateInfo::handleTypes does not include EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_BIT, a ExportFenceWin32HandleInfoKHR structure must not be included in the pNext chain of FenceCreateInfo

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_EXPORT_FENCE_WIN32_HANDLE_INFO_KHR

If pAttributes is not NULL, pAttributes must be a valid pointer to a valid SECURITY_ATTRIBUTES value

See Also

StructureType

Constructors

ExportFenceWin32HandleInfoKHR

Fields

attributes :: Ptr SECURITY_ATTRIBUTES

pAttributes is a pointer to a Windows SECURITY_ATTRIBUTES structure specifying security attributes of the handle.

dwAccess :: DWORD

dwAccess is a DWORD specifying access rights of the handle.

name :: LPCWSTR

name is a null-terminated UTF-16 string to associate with the underlying synchronization primitive referenced by NT handles exported from the created fence.

data FenceGetWin32HandleInfoKHR

VkFenceGetWin32HandleInfoKHR - Structure describing a Win32 handle fence export operation

Description

The properties of the handle returned depend on the value of handleType. See ExternalFenceHandleTypeFlagBits for a description of the properties of the defined external fence handle types.

Valid Usage

handleType must have been included in ExportFenceCreateInfo::handleTypes when the fence’s current payload was created

If handleType is defined as an NT handle, getFenceWin32HandleKHR must be called no more than once for each valid unique combination of fence and handleType

fence must not currently have its payload replaced by an imported payload as described below in Importing Fence Payloads unless that imported payload’s handle type was included in ExternalFenceProperties::exportFromImportedHandleTypes for handleType

If handleType refers to a handle type with copy payload transference semantics, fence must be signaled, or have an associated fence signal operation pending execution

handleType must be defined as an NT handle or a global share handle

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_FENCE_GET_WIN32_HANDLE_INFO_KHR

pNext must be NULL

fence must be a valid Fence handle

handleType must be a valid ExternalFenceHandleTypeFlagBits value

See Also

ExternalFenceHandleTypeFlagBits, Fence, StructureType, getFenceWin32HandleKHR

Constructors

FenceGetWin32HandleInfoKHR

Fields

fence :: Fence

fence is the fence from which state will be exported.

handleType :: ExternalFenceHandleTypeFlagBits

handleType is the type of handle requested.

type KHR_EXTERNAL_FENCE_WIN32_SPEC_VERSION = 1

pattern KHR_EXTERNAL_FENCE_WIN32_SPEC_VERSION :: forall a. Integral a => a

type KHR_EXTERNAL_FENCE_WIN32_EXTENSION_NAME = "VK_KHR_external_fence_win32"

pattern KHR_EXTERNAL_FENCE_WIN32_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

type HANDLE = Ptr ()

type DWORD = Word32

type LPCWSTR = Ptr CWchar

data SECURITY_ATTRIBUTES