Vulkan.Extensions.VK_KHR_external_semaphore_capabilities

Description

Name

VK_KHR_external_semaphore_capabilities - instance extension

VK_KHR_external_semaphore_capabilities

Name String

VK_KHR_external_semaphore_capabilities

Extension Type

Instance extension

Registered Extension Number

77

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_get_physical_device_properties2

Deprecation state

Promoted to Vulkan 1.1

Contact

James Jones https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_external_semaphore_capabilities:%20&body=@cubanismo%20

Other Extension Metadata

Last Modified Date

2016-10-20

IP Status

No known IP claims.

Interactions and External Dependencies

Promoted to Vulkan 1.1 Core

Contributors

Jesse Hall, Google

James Jones, NVIDIA

Jeff Juliano, NVIDIA

Description

An application may wish to reference device semaphores in multiple Vulkan logical devices or instances, in multiple processes, and/or in multiple APIs. This extension provides a set of capability queries and handle definitions that allow an application to determine what types of “external” semaphore handles an implementation supports for a given set of use cases.

Promotion to Vulkan 1.1

All functionality in this extension is included in core Vulkan 1.1, with the KHR suffix omitted. The original type, enum and command names are still available as aliases of the core functionality.

New Commands

getPhysicalDeviceExternalSemaphorePropertiesKHR

New Structures

ExternalSemaphorePropertiesKHR

PhysicalDeviceExternalSemaphoreInfoKHR

Extending PhysicalDeviceProperties2:

PhysicalDeviceIDPropertiesKHR

New Enums

ExternalSemaphoreFeatureFlagBitsKHR

ExternalSemaphoreHandleTypeFlagBitsKHR

New Bitmasks

ExternalSemaphoreFeatureFlagsKHR

ExternalSemaphoreHandleTypeFlagsKHR

New Enum Constants

KHR_EXTERNAL_SEMAPHORE_CAPABILITIES_EXTENSION_NAME

KHR_EXTERNAL_SEMAPHORE_CAPABILITIES_SPEC_VERSION

LUID_SIZE_KHR

Extending ExternalSemaphoreFeatureFlagBits:

EXTERNAL_SEMAPHORE_FEATURE_EXPORTABLE_BIT_KHR

EXTERNAL_SEMAPHORE_FEATURE_IMPORTABLE_BIT_KHR

Extending ExternalSemaphoreHandleTypeFlagBits:

EXTERNAL_SEMAPHORE_HANDLE_TYPE_D3D12_FENCE_BIT_KHR

EXTERNAL_SEMAPHORE_HANDLE_TYPE_OPAQUE_FD_BIT_KHR

EXTERNAL_SEMAPHORE_HANDLE_TYPE_OPAQUE_WIN32_BIT_KHR

EXTERNAL_SEMAPHORE_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_KHR

EXTERNAL_SEMAPHORE_HANDLE_TYPE_SYNC_FD_BIT_KHR

Extending StructureType:

STRUCTURE_TYPE_EXTERNAL_SEMAPHORE_PROPERTIES_KHR

STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_SEMAPHORE_INFO_KHR

STRUCTURE_TYPE_PHYSICAL_DEVICE_ID_PROPERTIES_KHR

Version History

Revision 1, 2016-10-20 (James Jones)

Initial revision

See Also

LUID_SIZE_KHR, ExternalSemaphoreFeatureFlagBitsKHR, ExternalSemaphoreFeatureFlagsKHR, ExternalSemaphoreHandleTypeFlagBitsKHR, ExternalSemaphoreHandleTypeFlagsKHR, ExternalSemaphorePropertiesKHR, PhysicalDeviceExternalSemaphoreInfoKHR, PhysicalDeviceIDPropertiesKHR, getPhysicalDeviceExternalSemaphorePropertiesKHR

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_EXTERNAL_SEMAPHORE_INFO_KHR :: StructureType

pattern STRUCTURE_TYPE_EXTERNAL_SEMAPHORE_PROPERTIES_KHR :: StructureType

pattern EXTERNAL_SEMAPHORE_HANDLE_TYPE_OPAQUE_FD_BIT_KHR :: ExternalSemaphoreHandleTypeFlagBits

pattern EXTERNAL_SEMAPHORE_HANDLE_TYPE_OPAQUE_WIN32_BIT_KHR :: ExternalSemaphoreHandleTypeFlagBits

pattern EXTERNAL_SEMAPHORE_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_KHR :: ExternalSemaphoreHandleTypeFlagBits

pattern EXTERNAL_SEMAPHORE_HANDLE_TYPE_D3D12_FENCE_BIT_KHR :: ExternalSemaphoreHandleTypeFlagBits

pattern EXTERNAL_SEMAPHORE_HANDLE_TYPE_SYNC_FD_BIT_KHR :: ExternalSemaphoreHandleTypeFlagBits

pattern EXTERNAL_SEMAPHORE_FEATURE_EXPORTABLE_BIT_KHR :: ExternalSemaphoreFeatureFlagBits

pattern EXTERNAL_SEMAPHORE_FEATURE_IMPORTABLE_BIT_KHR :: ExternalSemaphoreFeatureFlagBits

getPhysicalDeviceExternalSemaphorePropertiesKHR :: forall (a :: [Type]) io. (Extendss PhysicalDeviceExternalSemaphoreInfo a, PokeChain a, MonadIO io) => PhysicalDevice -> PhysicalDeviceExternalSemaphoreInfo a -> io ExternalSemaphoreProperties

type ExternalSemaphoreHandleTypeFlagsKHR = ExternalSemaphoreHandleTypeFlags

type ExternalSemaphoreFeatureFlagsKHR = ExternalSemaphoreFeatureFlags

type ExternalSemaphoreHandleTypeFlagBitsKHR = ExternalSemaphoreHandleTypeFlagBits

type ExternalSemaphoreFeatureFlagBitsKHR = ExternalSemaphoreFeatureFlagBits

type PhysicalDeviceExternalSemaphoreInfoKHR = PhysicalDeviceExternalSemaphoreInfo

type ExternalSemaphorePropertiesKHR = ExternalSemaphoreProperties

type KHR_EXTERNAL_SEMAPHORE_CAPABILITIES_SPEC_VERSION = 1

pattern KHR_EXTERNAL_SEMAPHORE_CAPABILITIES_SPEC_VERSION :: forall a. Integral a => a

type KHR_EXTERNAL_SEMAPHORE_CAPABILITIES_EXTENSION_NAME = "VK_KHR_external_semaphore_capabilities"

pattern KHR_EXTERNAL_SEMAPHORE_CAPABILITIES_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

type PhysicalDeviceIDPropertiesKHR = PhysicalDeviceIDProperties

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_ID_PROPERTIES_KHR :: StructureType

pattern LUID_SIZE_KHR :: Integral a => a