Vulkan.Extensions.VK_KHR_image_format_list

Description

Name

VK_KHR_image_format_list - device extension

VK_KHR_image_format_list

Name String

VK_KHR_image_format_list

Extension Type

Device extension

Registered Extension Number

148

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Deprecation state

Promoted to Vulkan 1.2

Contact

Jason Ekstrand https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_image_format_list:%20&body=@jekstrand%20

Other Extension Metadata

Last Modified Date

2017-03-20

Interactions and External Dependencies

Promoted to Vulkan 1.2 Core

IP Status

No known IP claims.

Contributors

Jason Ekstrand, Intel

Jan-Harald Fredriksen, ARM

Jeff Bolz, NVIDIA

Jeff Leger, Qualcomm

Neil Henning, Codeplay

Description

On some implementations, setting the IMAGE_CREATE_MUTABLE_FORMAT_BIT on image creation can cause access to that image to perform worse than an equivalent image created without IMAGE_CREATE_MUTABLE_FORMAT_BIT because the implementation does not know what view formats will be paired with the image.

This extension allows an application to provide the list of all formats that can be used with an image when it is created. The implementation may then be able to create a more efficient image that supports the subset of formats required by the application without having to support all formats in the format compatibility class of the image format.

Promotion to Vulkan 1.2

All functionality in this extension is included in core Vulkan 1.2, with the KHR suffix omitted. The original type, enum and command names are still available as aliases of the core functionality.

New Structures

Extending ImageCreateInfo, SwapchainCreateInfoKHR, PhysicalDeviceImageFormatInfo2:

ImageFormatListCreateInfoKHR

New Enum Constants

KHR_IMAGE_FORMAT_LIST_EXTENSION_NAME

KHR_IMAGE_FORMAT_LIST_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_IMAGE_FORMAT_LIST_CREATE_INFO_KHR

Version History

Revision 1, 2017-03-20 (Jason Ekstrand)

Initial revision

See Also

ImageFormatListCreateInfoKHR

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

pattern STRUCTURE_TYPE_IMAGE_FORMAT_LIST_CREATE_INFO_KHR :: StructureType

type ImageFormatListCreateInfoKHR = ImageFormatListCreateInfo

type KHR_IMAGE_FORMAT_LIST_SPEC_VERSION = 1

pattern KHR_IMAGE_FORMAT_LIST_SPEC_VERSION :: forall a. Integral a => a

type KHR_IMAGE_FORMAT_LIST_EXTENSION_NAME = "VK_KHR_image_format_list"

pattern KHR_IMAGE_FORMAT_LIST_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a