Vulkan.Extensions.VK_KHR_imageless_framebuffer

Description

Name

VK_KHR_imageless_framebuffer - device extension

VK_KHR_imageless_framebuffer

Name String

VK_KHR_imageless_framebuffer

Extension Type

Device extension

Registered Extension Number

109

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_maintenance2

Requires VK_KHR_image_format_list

Deprecation state

Promoted to Vulkan 1.2

Contact

Tobias Hector https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_imageless_framebuffer:%20&body=@tobias%20

Other Extension Metadata

Last Modified Date

2018-12-14

Interactions and External Dependencies

Promoted to Vulkan 1.2 Core

Contributors

Tobias Hector

Graham Wihlidal

Description

This extension allows framebuffers to be created without the need for creating images first, allowing more flexibility in how they are used, and avoiding the need for many of the confusing compatibility rules.

Framebuffers are now created with a small amount of additional metadata about the image views that will be used in FramebufferAttachmentsCreateInfoKHR, and the actual image views are provided at render pass begin time via RenderPassAttachmentBeginInfoKHR.

Promotion to Vulkan 1.2

All functionality in this extension is included in core Vulkan 1.2, with the KHR suffix omitted. The original type, enum and command names are still available as aliases of the core functionality.

New Structures

FramebufferAttachmentImageInfoKHR

Extending FramebufferCreateInfo:

FramebufferAttachmentsCreateInfoKHR

Extending PhysicalDeviceFeatures2, DeviceCreateInfo:

PhysicalDeviceImagelessFramebufferFeaturesKHR

Extending RenderPassBeginInfo:

RenderPassAttachmentBeginInfoKHR

New Enum Constants

KHR_IMAGELESS_FRAMEBUFFER_EXTENSION_NAME

KHR_IMAGELESS_FRAMEBUFFER_SPEC_VERSION

Extending FramebufferCreateFlagBits:

FRAMEBUFFER_CREATE_IMAGELESS_BIT_KHR

Extending StructureType:

STRUCTURE_TYPE_FRAMEBUFFER_ATTACHMENTS_CREATE_INFO_KHR

STRUCTURE_TYPE_FRAMEBUFFER_ATTACHMENT_IMAGE_INFO_KHR

STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGELESS_FRAMEBUFFER_FEATURES_KHR

STRUCTURE_TYPE_RENDER_PASS_ATTACHMENT_BEGIN_INFO_KHR

Version History

Revision 1, 2018-12-14 (Tobias Hector)

Internal revisions

See Also

FramebufferAttachmentImageInfoKHR, FramebufferAttachmentsCreateInfoKHR, PhysicalDeviceImagelessFramebufferFeaturesKHR, RenderPassAttachmentBeginInfoKHR

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_IMAGELESS_FRAMEBUFFER_FEATURES_KHR :: StructureType

pattern STRUCTURE_TYPE_FRAMEBUFFER_ATTACHMENTS_CREATE_INFO_KHR :: StructureType

pattern STRUCTURE_TYPE_FRAMEBUFFER_ATTACHMENT_IMAGE_INFO_KHR :: StructureType

pattern STRUCTURE_TYPE_RENDER_PASS_ATTACHMENT_BEGIN_INFO_KHR :: StructureType

pattern FRAMEBUFFER_CREATE_IMAGELESS_BIT_KHR :: FramebufferCreateFlagBits

type PhysicalDeviceImagelessFramebufferFeaturesKHR = PhysicalDeviceImagelessFramebufferFeatures

type FramebufferAttachmentsCreateInfoKHR = FramebufferAttachmentsCreateInfo

type FramebufferAttachmentImageInfoKHR = FramebufferAttachmentImageInfo

type RenderPassAttachmentBeginInfoKHR = RenderPassAttachmentBeginInfo

type KHR_IMAGELESS_FRAMEBUFFER_SPEC_VERSION = 1

pattern KHR_IMAGELESS_FRAMEBUFFER_SPEC_VERSION :: forall a. Integral a => a

type KHR_IMAGELESS_FRAMEBUFFER_EXTENSION_NAME = "VK_KHR_imageless_framebuffer"

pattern KHR_IMAGELESS_FRAMEBUFFER_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a