Vulkan.Extensions.VK_KHR_incremental_present

Documentation

data PresentRegionsKHR

VkPresentRegionsKHR - Structure hint of rectangular regions changed by vkQueuePresentKHR

Valid Usage

swapchainCount must be the same value as PresentInfoKHR::swapchainCount, where PresentInfoKHR is included in the pNext chain of this PresentRegionsKHR structure

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PRESENT_REGIONS_KHR

If pRegions is not NULL, pRegions must be a valid pointer to an array of swapchainCount valid PresentRegionKHR structures

swapchainCount must be greater than 0

See Also

PresentRegionKHR, StructureType

Constructors

PresentRegionsKHR

Fields

swapchainCount :: Word32

swapchainCount is the number of swapchains being presented to by this command.

regions :: Vector PresentRegionKHR

pRegions is NULL or a pointer to an array of PresentRegionKHR elements with swapchainCount entries. If not NULL, each element of pRegions contains the region that has changed since the last present to the swapchain in the corresponding entry in the PresentInfoKHR::pSwapchains array.

data PresentRegionKHR

VkPresentRegionKHR - Structure containing rectangular region changed by vkQueuePresentKHR for a given VkImage

Valid Usage (Implicit)

If rectangleCount is not 0, and pRectangles is not NULL, pRectangles must be a valid pointer to an array of rectangleCount valid RectLayerKHR structures

See Also

PresentRegionsKHR, RectLayerKHR

Constructors

PresentRegionKHR

Fields

rectangleCount :: Word32

rectangleCount is the number of rectangles in pRectangles, or zero if the entire image has changed and should be presented.

rectangles :: Vector RectLayerKHR

pRectangles is either NULL or a pointer to an array of RectLayerKHR structures. The RectLayerKHR structure is the framebuffer coordinates, plus layer, of a portion of a presentable image that has changed and must be presented. If non-NULL, each entry in pRectangles is a rectangle of the given image that has changed since the last image was presented to the given swapchain.

data RectLayerKHR

VkRectLayerKHR - Structure containing a rectangle, including layer, changed by vkQueuePresentKHR for a given VkImage

Valid Usage

The sum of offset and extent must be no greater than the imageExtent member of the SwapchainCreateInfoKHR structure passed to createSwapchainKHR

layer must be less than the imageArrayLayers member of the SwapchainCreateInfoKHR structure passed to createSwapchainKHR

Some platforms allow the size of a surface to change, and then scale the pixels of the image to fit the surface. RectLayerKHR specifies pixels of the swapchain’s image(s), which will be constant for the life of the swapchain.

See Also

Extent2D, Offset2D, PresentRegionKHR

Constructors

RectLayerKHR

Fields

offset :: Offset2D

offset is the origin of the rectangle, in pixels.

extent :: Extent2D

extent is the size of the rectangle, in pixels.

layer :: Word32

layer is the layer of the image. For images with only one layer, the value of layer must be 0.

type KHR_INCREMENTAL_PRESENT_SPEC_VERSION = 1

pattern KHR_INCREMENTAL_PRESENT_SPEC_VERSION :: forall a. Integral a => a

type KHR_INCREMENTAL_PRESENT_EXTENSION_NAME = "VK_KHR_incremental_present"

pattern KHR_INCREMENTAL_PRESENT_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a