Vulkan.Extensions.VK_KHR_pipeline_library

Description

Name

VK_KHR_pipeline_library - device extension

VK_KHR_pipeline_library

Name String

VK_KHR_pipeline_library

Extension Type

Device extension

Registered Extension Number

291

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Contact

Christoph Kubisch https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_pipeline_library:%20&body=@pixeljetstream%20

Other Extension Metadata

Last Modified Date

2020-01-08

IP Status

No known IP claims.

Contributors

See contributors to VK_KHR_ray_tracing

Description

A pipeline library is a special pipeline that cannot be bound, instead it defines a set of shaders and shader groups which can be linked into other pipelines. This extension defines the infrastructure for pipeline libraries, but does not specify the creation or usage of pipeline libraries. This is left to additional dependent extensions.

New Structures

PipelineLibraryCreateInfoKHR

New Enum Constants

KHR_PIPELINE_LIBRARY_EXTENSION_NAME

KHR_PIPELINE_LIBRARY_SPEC_VERSION

Extending PipelineCreateFlagBits:

PIPELINE_CREATE_LIBRARY_BIT_KHR

Extending StructureType:

STRUCTURE_TYPE_PIPELINE_LIBRARY_CREATE_INFO_KHR

Version History

Revision 1, 2020-01-08 (Christoph Kubisch)

Initial draft.

See Also

PipelineLibraryCreateInfoKHR

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data PipelineLibraryCreateInfoKHR

VkPipelineLibraryCreateInfoKHR - Structure specifying pipeline libraries to use when creating a pipeline

Valid Usage

Each element of pLibraries must have been created with PIPELINE_CREATE_LIBRARY_BIT_KHR

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PIPELINE_LIBRARY_CREATE_INFO_KHR

pNext must be NULL

If libraryCount is not 0, pLibraries must be a valid pointer to an array of libraryCount valid Pipeline handles

See Also

Pipeline, RayTracingPipelineCreateInfoKHR, StructureType

Constructors

PipelineLibraryCreateInfoKHR

Fields

libraries :: Vector Pipeline

pLibraries is an array of pipeline libraries to use when creating a pipeline.

type KHR_PIPELINE_LIBRARY_SPEC_VERSION = 1

pattern KHR_PIPELINE_LIBRARY_SPEC_VERSION :: forall a. Integral a => a

type KHR_PIPELINE_LIBRARY_EXTENSION_NAME = "VK_KHR_pipeline_library"

pattern KHR_PIPELINE_LIBRARY_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a