Vulkan.Extensions.VK_KHR_relaxed_block_layout

Description

Name

VK_KHR_relaxed_block_layout - device extension

VK_KHR_relaxed_block_layout

Name String

VK_KHR_relaxed_block_layout

Extension Type

Device extension

Registered Extension Number

145

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Deprecation state

Promoted to Vulkan 1.1

Contact

John Kessenich https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_relaxed_block_layout:%20&body=@johnkslang%20

Other Extension Metadata

Last Modified Date

2017-03-26

IP Status

No known IP claims.

Interactions and External Dependencies

Promoted to Vulkan 1.1 Core

Contributors

John Kessenich, Google

Description

The VK_KHR_relaxed_block_layout extension allows implementations to indicate they can support more variation in block Offset decorations. For example, placing a vector of three floats at an offset of 16×N + 4.

See Offset and Stride Assignment for details.

Promotion to Vulkan 1.1

All functionality in this extension is included in core Vulkan 1.1, with the KHR suffix omitted. The original type, enum and command names are still available as aliases of the core functionality.

New Enum Constants

KHR_RELAXED_BLOCK_LAYOUT_EXTENSION_NAME

KHR_RELAXED_BLOCK_LAYOUT_SPEC_VERSION

Version History

Revision 1, 2017-03-26 (JohnK)

See Also

No cross-references are available

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

type KHR_RELAXED_BLOCK_LAYOUT_SPEC_VERSION = 1

pattern KHR_RELAXED_BLOCK_LAYOUT_SPEC_VERSION :: forall a. Integral a => a

type KHR_RELAXED_BLOCK_LAYOUT_EXTENSION_NAME = "VK_KHR_relaxed_block_layout"

pattern KHR_RELAXED_BLOCK_LAYOUT_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a