Vulkan.Extensions.VK_KHR_sampler_ycbcr_conversion

Description

Name

VK_KHR_sampler_ycbcr_conversion - device extension

VK_KHR_sampler_ycbcr_conversion

Name String

VK_KHR_sampler_ycbcr_conversion

Extension Type

Device extension

Registered Extension Number

157

Revision

14

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_maintenance1

Requires VK_KHR_bind_memory2

Requires VK_KHR_get_memory_requirements2

Requires VK_KHR_get_physical_device_properties2

Deprecation state

Promoted to Vulkan 1.1

Contact

Andrew Garrard https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_sampler_ycbcr_conversion:%20&body=@fluppeteer%20

Other Extension Metadata

Last Modified Date

2017-08-11

IP Status

No known IP claims.

Interactions and External Dependencies

Promoted to Vulkan 1.1 Core

Contributors

Andrew Garrard, Samsung Electronics

Tobias Hector, Imagination Technologies

James Jones, NVIDIA

Daniel Koch, NVIDIA

Daniel Rakos, AMD

Romain Guy, Google

Jesse Hall, Google

Tom Cooksey, ARM Ltd

Jeff Leger, Qualcomm Technologies, Inc

Jan-Harald Fredriksen, ARM Ltd

Jan Outters, Samsung Electronics

Alon Or-bach, Samsung Electronics

Michael Worcester, Imagination Technologies

Jeff Bolz, NVIDIA

Tony Zlatinski, NVIDIA

Matthew Netsch, Qualcomm Technologies, Inc

Description

The use of Y′CBCR sampler conversion is an area in 3D graphics not used by most Vulkan developers. It’s mainly used for processing inputs from video decoders and cameras. The use of the extension assumes basic knowledge of Y′CBCR concepts.

This extension provides the ability to perform specified color space conversions during texture sampling operations for the Y′CBCR color space natively. It also adds a selection of multi-planar formats, image aspect plane, and the ability to bind memory to the planes of an image collectively or separately.

Promotion to Vulkan 1.1

All functionality in this extension is included in core Vulkan 1.1, with the KHR suffix omitted. However, if Vulkan 1.1 is supported and this extension is not, the samplerYcbcrConversion capability is optional. The original type, enum and command names are still available as aliases of the core functionality.

New Object Types

SamplerYcbcrConversionKHR

New Commands

createSamplerYcbcrConversionKHR

destroySamplerYcbcrConversionKHR

New Structures

SamplerYcbcrConversionCreateInfoKHR

Extending BindImageMemoryInfo:

BindImagePlaneMemoryInfoKHR

Extending ImageFormatProperties2:

SamplerYcbcrConversionImageFormatPropertiesKHR

Extending ImageMemoryRequirementsInfo2:

ImagePlaneMemoryRequirementsInfoKHR

Extending PhysicalDeviceFeatures2, DeviceCreateInfo:

PhysicalDeviceSamplerYcbcrConversionFeaturesKHR

Extending SamplerCreateInfo, ImageViewCreateInfo:

SamplerYcbcrConversionInfoKHR

New Enums

ChromaLocationKHR

SamplerYcbcrModelConversionKHR

SamplerYcbcrRangeKHR

New Enum Constants

KHR_SAMPLER_YCBCR_CONVERSION_EXTENSION_NAME

KHR_SAMPLER_YCBCR_CONVERSION_SPEC_VERSION

Extending ChromaLocation:

CHROMA_LOCATION_COSITED_EVEN_KHR

CHROMA_LOCATION_MIDPOINT_KHR

Extending DebugReportObjectTypeEXT:

DEBUG_REPORT_OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION_KHR_EXT

Extending Format:

FORMAT_B10X6G10X6R10X6G10X6_422_UNORM_4PACK16_KHR

FORMAT_B12X4G12X4R12X4G12X4_422_UNORM_4PACK16_KHR

FORMAT_B16G16R16G16_422_UNORM_KHR

FORMAT_B8G8R8G8_422_UNORM_KHR

FORMAT_G10X6B10X6G10X6R10X6_422_UNORM_4PACK16_KHR

FORMAT_G10X6_B10X6R10X6_2PLANE_420_UNORM_3PACK16_KHR

FORMAT_G10X6_B10X6R10X6_2PLANE_422_UNORM_3PACK16_KHR

FORMAT_G10X6_B10X6_R10X6_3PLANE_420_UNORM_3PACK16_KHR

FORMAT_G10X6_B10X6_R10X6_3PLANE_422_UNORM_3PACK16_KHR

FORMAT_G10X6_B10X6_R10X6_3PLANE_444_UNORM_3PACK16_KHR

FORMAT_G12X4B12X4G12X4R12X4_422_UNORM_4PACK16_KHR

FORMAT_G12X4_B12X4R12X4_2PLANE_420_UNORM_3PACK16_KHR

FORMAT_G12X4_B12X4R12X4_2PLANE_422_UNORM_3PACK16_KHR

FORMAT_G12X4_B12X4_R12X4_3PLANE_420_UNORM_3PACK16_KHR

FORMAT_G12X4_B12X4_R12X4_3PLANE_422_UNORM_3PACK16_KHR

FORMAT_G12X4_B12X4_R12X4_3PLANE_444_UNORM_3PACK16_KHR

FORMAT_G16B16G16R16_422_UNORM_KHR

FORMAT_G16_B16R16_2PLANE_420_UNORM_KHR

FORMAT_G16_B16R16_2PLANE_422_UNORM_KHR

FORMAT_G16_B16_R16_3PLANE_420_UNORM_KHR

FORMAT_G16_B16_R16_3PLANE_422_UNORM_KHR

FORMAT_G16_B16_R16_3PLANE_444_UNORM_KHR

FORMAT_G8B8G8R8_422_UNORM_KHR

FORMAT_G8_B8R8_2PLANE_420_UNORM_KHR

FORMAT_G8_B8R8_2PLANE_422_UNORM_KHR

FORMAT_G8_B8_R8_3PLANE_420_UNORM_KHR

FORMAT_G8_B8_R8_3PLANE_422_UNORM_KHR

FORMAT_G8_B8_R8_3PLANE_444_UNORM_KHR

FORMAT_R10X6G10X6B10X6A10X6_UNORM_4PACK16_KHR

FORMAT_R10X6G10X6_UNORM_2PACK16_KHR

FORMAT_R10X6_UNORM_PACK16_KHR

FORMAT_R12X4G12X4B12X4A12X4_UNORM_4PACK16_KHR

FORMAT_R12X4G12X4_UNORM_2PACK16_KHR

FORMAT_R12X4_UNORM_PACK16_KHR

Extending FormatFeatureFlagBits:

FORMAT_FEATURE_COSITED_CHROMA_SAMPLES_BIT_KHR

FORMAT_FEATURE_DISJOINT_BIT_KHR

FORMAT_FEATURE_MIDPOINT_CHROMA_SAMPLES_BIT_KHR

FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_BIT_KHR

FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_FORCEABLE_BIT_KHR

FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_LINEAR_FILTER_BIT_KHR

FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_SEPARATE_RECONSTRUCTION_FILTER_BIT_KHR

Extending ImageAspectFlagBits:

IMAGE_ASPECT_PLANE_0_BIT_KHR

IMAGE_ASPECT_PLANE_1_BIT_KHR

IMAGE_ASPECT_PLANE_2_BIT_KHR

Extending ImageCreateFlagBits:

IMAGE_CREATE_DISJOINT_BIT_KHR

Extending ObjectType:

OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION_KHR

Extending SamplerYcbcrModelConversion:

SAMPLER_YCBCR_MODEL_CONVERSION_RGB_IDENTITY_KHR

SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_2020_KHR

SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_601_KHR

SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_709_KHR

SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_IDENTITY_KHR

Extending SamplerYcbcrRange:

SAMPLER_YCBCR_RANGE_ITU_FULL_KHR

SAMPLER_YCBCR_RANGE_ITU_NARROW_KHR

Extending StructureType:

STRUCTURE_TYPE_BIND_IMAGE_PLANE_MEMORY_INFO_KHR

STRUCTURE_TYPE_IMAGE_PLANE_MEMORY_REQUIREMENTS_INFO_KHR

STRUCTURE_TYPE_PHYSICAL_DEVICE_SAMPLER_YCBCR_CONVERSION_FEATURES_KHR

STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_CREATE_INFO_KHR

STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_IMAGE_FORMAT_PROPERTIES_KHR

STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_INFO_KHR

If VK_EXT_debug_report is supported:

Extending DebugReportObjectTypeEXT:

DEBUG_REPORT_OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION_EXT

Version History

Revision 1, 2017-01-24 (Andrew Garrard)

Initial draft

Revision 2, 2017-01-25 (Andrew Garrard)

After initial feedback

Revision 3, 2017-01-27 (Andrew Garrard)

Higher bit depth formats, renaming, swizzle

Revision 4, 2017-02-22 (Andrew Garrard)

Added query function, formats as RGB, clarifications

Revision 5, 2017-04 (Andrew Garrard)

Simplified query and removed output conversions

Revision 6, 2017-4-24 (Andrew Garrard)

Tidying, incorporated new image query, restored transfer functions

Revision 7, 2017-04-25 (Andrew Garrard)

Added cosited option/midpoint requirement for formats, "bypassConversion"

Revision 8, 2017-04-25 (Andrew Garrard)

Simplified further

Revision 9, 2017-04-27 (Andrew Garrard)

Disjoint no more

Revision 10, 2017-04-28 (Andrew Garrard)

Restored disjoint

Revision 11, 2017-04-29 (Andrew Garrard)

Now Ycbcr conversion, and KHR

Revision 12, 2017-06-06 (Andrew Garrard)

Added conversion to image view creation

Revision 13, 2017-07-13 (Andrew Garrard)

Allowed cosited-only chroma samples for formats

Revision 14, 2017-08-11 (Andrew Garrard)

Reflected quantization changes in BT.2100-1

See Also

BindImagePlaneMemoryInfoKHR, ChromaLocationKHR, ImagePlaneMemoryRequirementsInfoKHR, PhysicalDeviceSamplerYcbcrConversionFeaturesKHR, SamplerYcbcrConversionCreateInfoKHR, SamplerYcbcrConversionImageFormatPropertiesKHR, SamplerYcbcrConversionInfoKHR, SamplerYcbcrConversionKHR, SamplerYcbcrModelConversionKHR, SamplerYcbcrRangeKHR, createSamplerYcbcrConversionKHR, destroySamplerYcbcrConversionKHR

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

pattern STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_CREATE_INFO_KHR :: StructureType

pattern STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_INFO_KHR :: StructureType

pattern STRUCTURE_TYPE_BIND_IMAGE_PLANE_MEMORY_INFO_KHR :: StructureType

pattern STRUCTURE_TYPE_IMAGE_PLANE_MEMORY_REQUIREMENTS_INFO_KHR :: StructureType

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SAMPLER_YCBCR_CONVERSION_FEATURES_KHR :: StructureType

pattern STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_IMAGE_FORMAT_PROPERTIES_KHR :: StructureType

pattern DEBUG_REPORT_OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION_KHR_EXT :: DebugReportObjectTypeEXT

pattern OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION_KHR :: ObjectType

pattern FORMAT_G8B8G8R8_422_UNORM_KHR :: Format

pattern FORMAT_B8G8R8G8_422_UNORM_KHR :: Format

pattern FORMAT_G8_B8_R8_3PLANE_420_UNORM_KHR :: Format

pattern FORMAT_G8_B8R8_2PLANE_420_UNORM_KHR :: Format

pattern FORMAT_G8_B8_R8_3PLANE_422_UNORM_KHR :: Format

pattern FORMAT_G8_B8R8_2PLANE_422_UNORM_KHR :: Format

pattern FORMAT_G8_B8_R8_3PLANE_444_UNORM_KHR :: Format

pattern FORMAT_R10X6_UNORM_PACK16_KHR :: Format

pattern FORMAT_R10X6G10X6_UNORM_2PACK16_KHR :: Format

pattern FORMAT_R10X6G10X6B10X6A10X6_UNORM_4PACK16_KHR :: Format

pattern FORMAT_G10X6B10X6G10X6R10X6_422_UNORM_4PACK16_KHR :: Format

pattern FORMAT_B10X6G10X6R10X6G10X6_422_UNORM_4PACK16_KHR :: Format

pattern FORMAT_G10X6_B10X6_R10X6_3PLANE_420_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G10X6_B10X6R10X6_2PLANE_420_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G10X6_B10X6_R10X6_3PLANE_422_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G10X6_B10X6R10X6_2PLANE_422_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G10X6_B10X6_R10X6_3PLANE_444_UNORM_3PACK16_KHR :: Format

pattern FORMAT_R12X4_UNORM_PACK16_KHR :: Format

pattern FORMAT_R12X4G12X4_UNORM_2PACK16_KHR :: Format

pattern FORMAT_R12X4G12X4B12X4A12X4_UNORM_4PACK16_KHR :: Format

pattern FORMAT_G12X4B12X4G12X4R12X4_422_UNORM_4PACK16_KHR :: Format

pattern FORMAT_B12X4G12X4R12X4G12X4_422_UNORM_4PACK16_KHR :: Format

pattern FORMAT_G12X4_B12X4_R12X4_3PLANE_420_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G12X4_B12X4R12X4_2PLANE_420_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G12X4_B12X4_R12X4_3PLANE_422_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G12X4_B12X4R12X4_2PLANE_422_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G12X4_B12X4_R12X4_3PLANE_444_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G16B16G16R16_422_UNORM_KHR :: Format

pattern FORMAT_B16G16R16G16_422_UNORM_KHR :: Format

pattern FORMAT_G16_B16_R16_3PLANE_420_UNORM_KHR :: Format

pattern FORMAT_G16_B16R16_2PLANE_420_UNORM_KHR :: Format

pattern FORMAT_G16_B16_R16_3PLANE_422_UNORM_KHR :: Format

pattern FORMAT_G16_B16R16_2PLANE_422_UNORM_KHR :: Format

pattern FORMAT_G16_B16_R16_3PLANE_444_UNORM_KHR :: Format

pattern IMAGE_ASPECT_PLANE_0_BIT_KHR :: ImageAspectFlagBits

pattern IMAGE_ASPECT_PLANE_1_BIT_KHR :: ImageAspectFlagBits

pattern IMAGE_ASPECT_PLANE_2_BIT_KHR :: ImageAspectFlagBits

pattern IMAGE_CREATE_DISJOINT_BIT_KHR :: ImageCreateFlagBits

pattern FORMAT_FEATURE_MIDPOINT_CHROMA_SAMPLES_BIT_KHR :: FormatFeatureFlagBits

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_LINEAR_FILTER_BIT_KHR :: FormatFeatureFlagBits

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_SEPARATE_RECONSTRUCTION_FILTER_BIT_KHR :: FormatFeatureFlagBits

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_BIT_KHR :: FormatFeatureFlagBits

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_FORCEABLE_BIT_KHR :: FormatFeatureFlagBits

pattern FORMAT_FEATURE_DISJOINT_BIT_KHR :: FormatFeatureFlagBits

pattern FORMAT_FEATURE_COSITED_CHROMA_SAMPLES_BIT_KHR :: FormatFeatureFlagBits

pattern SAMPLER_YCBCR_MODEL_CONVERSION_RGB_IDENTITY_KHR :: SamplerYcbcrModelConversion

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_IDENTITY_KHR :: SamplerYcbcrModelConversion

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_709_KHR :: SamplerYcbcrModelConversion

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_601_KHR :: SamplerYcbcrModelConversion

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_2020_KHR :: SamplerYcbcrModelConversion

pattern SAMPLER_YCBCR_RANGE_ITU_FULL_KHR :: SamplerYcbcrRange

pattern SAMPLER_YCBCR_RANGE_ITU_NARROW_KHR :: SamplerYcbcrRange

pattern CHROMA_LOCATION_COSITED_EVEN_KHR :: ChromaLocation

pattern CHROMA_LOCATION_MIDPOINT_KHR :: ChromaLocation

createSamplerYcbcrConversionKHR :: forall (a :: [Type]) io. (Extendss SamplerYcbcrConversionCreateInfo a, PokeChain a, MonadIO io) => Device -> SamplerYcbcrConversionCreateInfo a -> ("allocator" ::: Maybe AllocationCallbacks) -> io SamplerYcbcrConversion

destroySamplerYcbcrConversionKHR :: MonadIO io => Device -> SamplerYcbcrConversion -> ("allocator" ::: Maybe AllocationCallbacks) -> io ()

type SamplerYcbcrConversionKHR = SamplerYcbcrConversion

type SamplerYcbcrModelConversionKHR = SamplerYcbcrModelConversion

type SamplerYcbcrRangeKHR = SamplerYcbcrRange

type ChromaLocationKHR = ChromaLocation

type SamplerYcbcrConversionInfoKHR = SamplerYcbcrConversionInfo

type SamplerYcbcrConversionCreateInfoKHR = SamplerYcbcrConversionCreateInfo

type BindImagePlaneMemoryInfoKHR = BindImagePlaneMemoryInfo

type ImagePlaneMemoryRequirementsInfoKHR = ImagePlaneMemoryRequirementsInfo

type PhysicalDeviceSamplerYcbcrConversionFeaturesKHR = PhysicalDeviceSamplerYcbcrConversionFeatures

type SamplerYcbcrConversionImageFormatPropertiesKHR = SamplerYcbcrConversionImageFormatProperties

type KHR_SAMPLER_YCBCR_CONVERSION_SPEC_VERSION = 14

pattern KHR_SAMPLER_YCBCR_CONVERSION_SPEC_VERSION :: forall a. Integral a => a

type KHR_SAMPLER_YCBCR_CONVERSION_EXTENSION_NAME = "VK_KHR_sampler_ycbcr_conversion"

pattern KHR_SAMPLER_YCBCR_CONVERSION_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype DebugReportObjectTypeEXT

VkDebugReportObjectTypeEXT - Specify the type of an object handle

Description

'

DebugReportObjectTypeEXT 	 Vulkan Handle Type

DEBUG_REPORT_OBJECT_TYPE_UNKNOWN_EXT 	 Unknown/Undefined Handle

DEBUG_REPORT_OBJECT_TYPE_INSTANCE_EXT 	 Instance

DEBUG_REPORT_OBJECT_TYPE_PHYSICAL_DEVICE_EXT 	 PhysicalDevice

DEBUG_REPORT_OBJECT_TYPE_DEVICE_EXT 	 Device

DEBUG_REPORT_OBJECT_TYPE_QUEUE_EXT 	 Queue

DEBUG_REPORT_OBJECT_TYPE_SEMAPHORE_EXT 	 Semaphore

DEBUG_REPORT_OBJECT_TYPE_COMMAND_BUFFER_EXT 	 CommandBuffer

DEBUG_REPORT_OBJECT_TYPE_FENCE_EXT 	 Fence

DEBUG_REPORT_OBJECT_TYPE_DEVICE_MEMORY_EXT 	 DeviceMemory

DEBUG_REPORT_OBJECT_TYPE_BUFFER_EXT 	 Buffer

DEBUG_REPORT_OBJECT_TYPE_IMAGE_EXT 	 Image

DEBUG_REPORT_OBJECT_TYPE_EVENT_EXT 	 Event

DEBUG_REPORT_OBJECT_TYPE_QUERY_POOL_EXT 	 QueryPool

DEBUG_REPORT_OBJECT_TYPE_BUFFER_VIEW_EXT 	 BufferView

DEBUG_REPORT_OBJECT_TYPE_IMAGE_VIEW_EXT 	 ImageView

DEBUG_REPORT_OBJECT_TYPE_SHADER_MODULE_EXT 	 ShaderModule

DEBUG_REPORT_OBJECT_TYPE_PIPELINE_CACHE_EXT 	 PipelineCache

DEBUG_REPORT_OBJECT_TYPE_PIPELINE_LAYOUT_EXT 	 PipelineLayout

DEBUG_REPORT_OBJECT_TYPE_RENDER_PASS_EXT 	 RenderPass

DEBUG_REPORT_OBJECT_TYPE_PIPELINE_EXT 	 Pipeline

DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_LAYOUT_EXT 	 DescriptorSetLayout

DEBUG_REPORT_OBJECT_TYPE_SAMPLER_EXT 	 Sampler

DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_POOL_EXT 	 DescriptorPool

DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_EXT 	 DescriptorSet

DEBUG_REPORT_OBJECT_TYPE_FRAMEBUFFER_EXT 	 Framebuffer

DEBUG_REPORT_OBJECT_TYPE_COMMAND_POOL_EXT 	 CommandPool

DEBUG_REPORT_OBJECT_TYPE_SURFACE_KHR_EXT 	 SurfaceKHR

DEBUG_REPORT_OBJECT_TYPE_SWAPCHAIN_KHR_EXT 	 SwapchainKHR

DEBUG_REPORT_OBJECT_TYPE_DEBUG_REPORT_CALLBACK_EXT_EXT 	 DebugReportCallbackEXT

DEBUG_REPORT_OBJECT_TYPE_DISPLAY_KHR_EXT 	 DisplayKHR

DEBUG_REPORT_OBJECT_TYPE_DISPLAY_MODE_KHR_EXT 	 DisplayModeKHR

DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_EXT 	 DescriptorUpdateTemplate

DebugReportObjectTypeEXT and Vulkan Handle Relationship

Note

The primary expected use of ERROR_VALIDATION_FAILED_EXT is for validation layer testing. It is not expected that an application would see this error code during normal use of the validation layers.

See Also

DebugMarkerObjectNameInfoEXT, DebugMarkerObjectTagInfoEXT, debugReportMessageEXT

Constructors

DebugReportObjectTypeEXT Int32

Bundled Patterns

pattern DEBUG_REPORT_OBJECT_TYPE_UNKNOWN_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_INSTANCE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_PHYSICAL_DEVICE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DEVICE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_QUEUE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SEMAPHORE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_COMMAND_BUFFER_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_FENCE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DEVICE_MEMORY_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_BUFFER_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_IMAGE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_EVENT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_QUERY_POOL_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_BUFFER_VIEW_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_IMAGE_VIEW_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SHADER_MODULE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_PIPELINE_CACHE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_PIPELINE_LAYOUT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_RENDER_PASS_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_PIPELINE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_LAYOUT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SAMPLER_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_POOL_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_FRAMEBUFFER_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_COMMAND_POOL_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SURFACE_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SWAPCHAIN_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DEBUG_REPORT_CALLBACK_EXT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DISPLAY_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DISPLAY_MODE_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_VALIDATION_CACHE_EXT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_ACCELERATION_STRUCTURE_NV_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_ACCELERATION_STRUCTURE_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_EXT :: DebugReportObjectTypeEXT