Vulkan.Extensions.VK_KHR_separate_depth_stencil_layouts

Description

Name

VK_KHR_separate_depth_stencil_layouts - device extension

VK_KHR_separate_depth_stencil_layouts

Name String

VK_KHR_separate_depth_stencil_layouts

Extension Type

Device extension

Registered Extension Number

242

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_get_physical_device_properties2

Requires VK_KHR_create_renderpass2

Deprecation state

Promoted to Vulkan 1.2

Contact

Piers Daniell https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_separate_depth_stencil_layouts:%20&body=@pdaniell-nv%20

Other Extension Metadata

Last Modified Date

2019-06-25

Interactions and External Dependencies

Promoted to Vulkan 1.2 Core

Contributors

Daniel Koch, NVIDIA

Jeff Bolz, NVIDIA

Jesse Barker, Unity

Tobias Hector, AMD

Description

This extension allows image memory barriers for depth/stencil images to have just one of the IMAGE_ASPECT_DEPTH_BIT or IMAGE_ASPECT_STENCIL_BIT aspect bits set, rather than require both. This allows their layouts to be set independently. To support depth/stencil images with different layouts for the depth and stencil aspects, the depth/stencil attachment interface has been updated to support a separate layout for stencil.

Promotion to Vulkan 1.2

All functionality in this extension is included in core Vulkan 1.2, with the KHR suffix omitted. The original type, enum and command names are still available as aliases of the core functionality.

New Structures

Extending AttachmentDescription2:

AttachmentDescriptionStencilLayoutKHR

Extending AttachmentReference2:

AttachmentReferenceStencilLayoutKHR

Extending PhysicalDeviceFeatures2, DeviceCreateInfo:

PhysicalDeviceSeparateDepthStencilLayoutsFeaturesKHR

New Enum Constants

KHR_SEPARATE_DEPTH_STENCIL_LAYOUTS_EXTENSION_NAME

KHR_SEPARATE_DEPTH_STENCIL_LAYOUTS_SPEC_VERSION

Extending ImageLayout:

IMAGE_LAYOUT_DEPTH_ATTACHMENT_OPTIMAL_KHR

IMAGE_LAYOUT_DEPTH_READ_ONLY_OPTIMAL_KHR

IMAGE_LAYOUT_STENCIL_ATTACHMENT_OPTIMAL_KHR

IMAGE_LAYOUT_STENCIL_READ_ONLY_OPTIMAL_KHR

Extending StructureType:

STRUCTURE_TYPE_ATTACHMENT_DESCRIPTION_STENCIL_LAYOUT_KHR

STRUCTURE_TYPE_ATTACHMENT_REFERENCE_STENCIL_LAYOUT_KHR

STRUCTURE_TYPE_PHYSICAL_DEVICE_SEPARATE_DEPTH_STENCIL_LAYOUTS_FEATURES_KHR

Version History

Revision 1, 2019-06-25 (Piers Daniell)

Internal revisions

See Also

AttachmentDescriptionStencilLayoutKHR, AttachmentReferenceStencilLayoutKHR, PhysicalDeviceSeparateDepthStencilLayoutsFeaturesKHR

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SEPARATE_DEPTH_STENCIL_LAYOUTS_FEATURES_KHR :: StructureType

pattern STRUCTURE_TYPE_ATTACHMENT_REFERENCE_STENCIL_LAYOUT_KHR :: StructureType

pattern STRUCTURE_TYPE_ATTACHMENT_DESCRIPTION_STENCIL_LAYOUT_KHR :: StructureType

pattern IMAGE_LAYOUT_DEPTH_ATTACHMENT_OPTIMAL_KHR :: ImageLayout

pattern IMAGE_LAYOUT_DEPTH_READ_ONLY_OPTIMAL_KHR :: ImageLayout

pattern IMAGE_LAYOUT_STENCIL_ATTACHMENT_OPTIMAL_KHR :: ImageLayout

pattern IMAGE_LAYOUT_STENCIL_READ_ONLY_OPTIMAL_KHR :: ImageLayout

type PhysicalDeviceSeparateDepthStencilLayoutsFeaturesKHR = PhysicalDeviceSeparateDepthStencilLayoutsFeatures

type AttachmentReferenceStencilLayoutKHR = AttachmentReferenceStencilLayout

type AttachmentDescriptionStencilLayoutKHR = AttachmentDescriptionStencilLayout

type KHR_SEPARATE_DEPTH_STENCIL_LAYOUTS_SPEC_VERSION = 1

pattern KHR_SEPARATE_DEPTH_STENCIL_LAYOUTS_SPEC_VERSION :: forall a. Integral a => a

type KHR_SEPARATE_DEPTH_STENCIL_LAYOUTS_EXTENSION_NAME = "VK_KHR_separate_depth_stencil_layouts"

pattern KHR_SEPARATE_DEPTH_STENCIL_LAYOUTS_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a