Vulkan.Extensions.VK_KHR_shader_atomic_int64

Description

Name

VK_KHR_shader_atomic_int64 - device extension

VK_KHR_shader_atomic_int64

Name String

VK_KHR_shader_atomic_int64

Extension Type

Device extension

Registered Extension Number

181

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_get_physical_device_properties2

Deprecation state

Promoted to Vulkan 1.2

Contact

Aaron Hagan https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_shader_atomic_int64:%20&body=@ahagan%20

Other Extension Metadata

Last Modified Date

2018-07-05

Interactions and External Dependencies

Promoted to Vulkan 1.2 Core

This extension enables GL_ARB_gpu_shader_int64 and GL_EXT_shader_atomic_int64 for GLSL source languages.

Contributors

Aaron Hagan, AMD

Daniel Rakos, AMD

Jeff Bolz, NVIDIA

Neil Henning, Codeplay

Description

This extension advertises the SPIR-V Int64Atomics capability for Vulkan, which allows a shader to contain 64-bit atomic operations on signed and unsigned integers. The supported operations include OpAtomicMin, OpAtomicMax, OpAtomicAnd, OpAtomicOr, OpAtomicXor, OpAtomicAdd, OpAtomicExchange, and OpAtomicCompareExchange.

Promotion to Vulkan 1.2

All functionality in this extension is included in core Vulkan 1.2, with the KHR suffix omitted. However, if Vulkan 1.2 is supported and this extension is not, the shaderBufferInt64Atomics capability is optional. The original type, enum and command names are still available as aliases of the core functionality.

New Structures

Extending PhysicalDeviceFeatures2, DeviceCreateInfo:

PhysicalDeviceShaderAtomicInt64FeaturesKHR

New Enum Constants

KHR_SHADER_ATOMIC_INT64_EXTENSION_NAME

KHR_SHADER_ATOMIC_INT64_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_ATOMIC_INT64_FEATURES_KHR

New SPIR-V Capabilities

Int64Atomics

Version History

Revision 1, 2018-07-05 (Aaron Hagan)

Internal revisions

See Also

PhysicalDeviceShaderAtomicInt64FeaturesKHR

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_ATOMIC_INT64_FEATURES_KHR :: StructureType

type PhysicalDeviceShaderAtomicInt64FeaturesKHR = PhysicalDeviceShaderAtomicInt64Features

type KHR_SHADER_ATOMIC_INT64_SPEC_VERSION = 1

pattern KHR_SHADER_ATOMIC_INT64_SPEC_VERSION :: forall a. Integral a => a

type KHR_SHADER_ATOMIC_INT64_EXTENSION_NAME = "VK_KHR_shader_atomic_int64"

pattern KHR_SHADER_ATOMIC_INT64_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a