Vulkan.Extensions.VK_KHR_shader_terminate_invocation

Description

Name

VK_KHR_shader_terminate_invocation - device extension

VK_KHR_shader_terminate_invocation

Name String

VK_KHR_shader_terminate_invocation

Extension Type

Device extension

Registered Extension Number

216

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_get_physical_device_properties2

Contact

Jesse Hall https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_shader_terminate_invocation:%20&body=@critsec%20

Last Modified Date

2020-08-11

IP Status

No known IP claims.

Interactions and External Dependencies

Requires the SPV_KHR_terminate_invocation SPIR-V extension.

Contributors

Alan Baker, Google

Jeff Bolz, NVIDIA

Jesse Hall, Google

Ralph Potter, Samsung

Tom Olson, Arm

Description

This extension adds Vulkan support for the SPV_KHR_terminate_invocation SPIR-V extension. That SPIR-V extension provides a new instruction, OpTerminateInvocation, which causes a shader invocation to immediately terminate and sets the coverage of shaded samples to 0; only previously executed instructions will have observable effects. The OpTerminateInvocation instruction, along with the OpDemoteToHelperInvocation instruction from the VK_EXT_shader_demote_to_helper_invocation extension, together replace the OpKill instruction, which could behave like either of these instructions. OpTerminateInvocation provides the behavior required by the GLSL discard statement, and should be used when available by GLSL compilers and applications that need the GLSL discard behavior.

New Structures

Extending PhysicalDeviceFeatures2, DeviceCreateInfo:

PhysicalDeviceShaderTerminateInvocationFeaturesKHR

New Enum Constants

KHR_SHADER_TERMINATE_INVOCATION_EXTENSION_NAME

KHR_SHADER_TERMINATE_INVOCATION_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADER_TERMINATE_INVOCATION_FEATURES_KHR

Version History

Revision 1, 2020-08-11 (Jesse Hall)

See Also

PhysicalDeviceShaderTerminateInvocationFeaturesKHR

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data PhysicalDeviceShaderTerminateInvocationFeaturesKHR

VkPhysicalDeviceShaderTerminateInvocationFeaturesKHR - Structure describing support for the SPIR-V SPV_KHR_terminate_invocation extension

Members

The members of the PhysicalDeviceShaderTerminateInvocationFeaturesKHR structure describe the following feature:

Description

If the PhysicalDeviceShaderTerminateInvocationFeaturesKHR structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with a value indicating whether the feature is supported. PhysicalDeviceShaderTerminateInvocationFeaturesKHR can also be included in the pNext chain of DeviceCreateInfo to enable the features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceShaderTerminateInvocationFeaturesKHR

Fields

shaderTerminateInvocation :: Bool

shaderTerminateInvocation specifies whether the implementation supports SPIR-V modules that use the SPV_KHR_terminate_invocation extension.

type KHR_SHADER_TERMINATE_INVOCATION_SPEC_VERSION = 1

pattern KHR_SHADER_TERMINATE_INVOCATION_SPEC_VERSION :: forall a. Integral a => a

type KHR_SHADER_TERMINATE_INVOCATION_EXTENSION_NAME = "VK_KHR_shader_terminate_invocation"

pattern KHR_SHADER_TERMINATE_INVOCATION_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a