Vulkan.Extensions.VK_KHR_storage_buffer_storage_class

Description

Name

VK_KHR_storage_buffer_storage_class - device extension

VK_KHR_storage_buffer_storage_class

Name String

VK_KHR_storage_buffer_storage_class

Extension Type

Device extension

Registered Extension Number

132

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Deprecation state

Promoted to Vulkan 1.1

Contact

Alexander Galazin https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_storage_buffer_storage_class:%20&body=@alegal-arm%20

Other Extension Metadata

Last Modified Date

2017-09-05

IP Status

No known IP claims.

Interactions and External Dependencies

This extension requires SPV_KHR_storage_buffer_storage_class

Promoted to Vulkan 1.1 Core

Contributors

Alexander Galazin, ARM

David Neto, Google

Description

This extension adds support for the following SPIR-V extension in Vulkan:

SPV_KHR_storage_buffer_storage_class

This extension provides a new SPIR-V StorageBuffer storage class. A Block-decorated object in this class is equivalent to a BufferBlock-decorated object in the Uniform storage class.

Promotion to Vulkan 1.1

All functionality in this extension is included in core Vulkan 1.1.

New Enum Constants

KHR_STORAGE_BUFFER_STORAGE_CLASS_EXTENSION_NAME

KHR_STORAGE_BUFFER_STORAGE_CLASS_SPEC_VERSION

Version History

Revision 1, 2017-03-23 (Alexander Galazin)

Initial draft

See Also

No cross-references are available

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

type KHR_STORAGE_BUFFER_STORAGE_CLASS_SPEC_VERSION = 1

pattern KHR_STORAGE_BUFFER_STORAGE_CLASS_SPEC_VERSION :: forall a. Integral a => a

type KHR_STORAGE_BUFFER_STORAGE_CLASS_EXTENSION_NAME = "VK_KHR_storage_buffer_storage_class"

pattern KHR_STORAGE_BUFFER_STORAGE_CLASS_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a