Vulkan.Extensions.VK_KHR_surface_protected_capabilities

Documentation

data SurfaceProtectedCapabilitiesKHR

VkSurfaceProtectedCapabilitiesKHR - Structure describing capability of a surface to be protected

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

SurfaceProtectedCapabilitiesKHR

Fields

supportsProtected :: Bool

supportsProtected specifies whether a protected swapchain created from PhysicalDeviceSurfaceInfo2KHR::surface for a particular windowing system can be displayed on screen or not. If supportsProtected is TRUE, then creation of swapchains with the SWAPCHAIN_CREATE_PROTECTED_BIT_KHR flag set must be supported for surface.

type KHR_SURFACE_PROTECTED_CAPABILITIES_SPEC_VERSION = 1

pattern KHR_SURFACE_PROTECTED_CAPABILITIES_SPEC_VERSION :: forall a. Integral a => a

type KHR_SURFACE_PROTECTED_CAPABILITIES_EXTENSION_NAME = "VK_KHR_surface_protected_capabilities"

pattern KHR_SURFACE_PROTECTED_CAPABILITIES_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a