Vulkan.Extensions.VK_KHR_vulkan_memory_model

Description

Name

VK_KHR_vulkan_memory_model - device extension

VK_KHR_vulkan_memory_model

Name String

VK_KHR_vulkan_memory_model

Extension Type

Device extension

Registered Extension Number

212

Revision

3

Extension and Version Dependencies

Requires Vulkan 1.0

Deprecation state

Promoted to Vulkan 1.2

Contact

Jeff Bolz https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_vulkan_memory_model:%20&body=@jeffbolznv%20

Other Extension Metadata

Last Modified Date

2018-12-10

IP Status

No known IP claims.

Interactions and External Dependencies

Promoted to Vulkan 1.2 Core

This extension requires SPV_KHR_vulkan_memory_model

Contributors

Jeff Bolz, NVIDIA

Alan Baker, Google

Tobias Hector, AMD

David Neto, Google

Robert Simpson, Qualcomm Technologies, Inc.

Brian Sumner, AMD

Description

The VK_KHR_vulkan_memory_model extension allows use of the Vulkan Memory Model, which formally defines how to synchronize memory accesses to the same memory locations performed by multiple shader invocations.

Note

Version 3 of the spec added a member (vulkanMemoryModelAvailabilityVisibilityChains) to PhysicalDeviceVulkanMemoryModelFeaturesKHR, which is an incompatible change from version 2.

Promotion to Vulkan 1.2

All functionality in this extension is included in core Vulkan 1.2, with the KHR suffix omitted. However, if Vulkan 1.2 is supported and this extension is not, the vulkanMemoryModel capability is optional. The original type, enum and command names are still available as aliases of the core functionality.

New Structures

Extending PhysicalDeviceFeatures2, DeviceCreateInfo:

PhysicalDeviceVulkanMemoryModelFeaturesKHR

New Enum Constants

KHR_VULKAN_MEMORY_MODEL_EXTENSION_NAME

KHR_VULKAN_MEMORY_MODEL_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_PHYSICAL_DEVICE_VULKAN_MEMORY_MODEL_FEATURES_KHR

New SPIR-V Capabilities

VulkanMemoryModelKHR

Version History

Revision 1, 2018-06-24 (Jeff Bolz)

Initial draft

Revision 3, 2018-12-10 (Jeff Bolz)

Add vulkanMemoryModelAvailabilityVisibilityChains member to the VkPhysicalDeviceVulkanMemoryModelFeaturesKHR structure.

See Also

PhysicalDeviceVulkanMemoryModelFeaturesKHR

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_VULKAN_MEMORY_MODEL_FEATURES_KHR :: StructureType

type PhysicalDeviceVulkanMemoryModelFeaturesKHR = PhysicalDeviceVulkanMemoryModelFeatures

type KHR_VULKAN_MEMORY_MODEL_SPEC_VERSION = 3

pattern KHR_VULKAN_MEMORY_MODEL_SPEC_VERSION :: forall a. Integral a => a

type KHR_VULKAN_MEMORY_MODEL_EXTENSION_NAME = "VK_KHR_vulkan_memory_model"

pattern KHR_VULKAN_MEMORY_MODEL_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a