Vulkan.Extensions.VK_KHR_win32_keyed_mutex

Description

Name

VK_KHR_win32_keyed_mutex - device extension

VK_KHR_win32_keyed_mutex

Name String

VK_KHR_win32_keyed_mutex

Extension Type

Device extension

Registered Extension Number

76

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_external_memory_win32

Contact

Carsten Rohde https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_KHR_win32_keyed_mutex:%20&body=@crohde%20

Other Extension Metadata

Last Modified Date

2016-10-21

IP Status

No known IP claims.

Contributors

James Jones, NVIDIA

Jeff Juliano, NVIDIA

Carsten Rohde, NVIDIA

Description

Applications that wish to import Direct3D 11 memory objects into the Vulkan API may wish to use the native keyed mutex mechanism to synchronize access to the memory between Vulkan and Direct3D. This extension provides a way for an application to access the keyed mutex associated with an imported Vulkan memory object when submitting command buffers to a queue.

New Structures

Extending SubmitInfo:

Win32KeyedMutexAcquireReleaseInfoKHR

New Enum Constants

KHR_WIN32_KEYED_MUTEX_EXTENSION_NAME

KHR_WIN32_KEYED_MUTEX_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_WIN32_KEYED_MUTEX_ACQUIRE_RELEASE_INFO_KHR

Version History

Revision 1, 2016-10-21 (James Jones)

Initial revision

See Also

Win32KeyedMutexAcquireReleaseInfoKHR

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data Win32KeyedMutexAcquireReleaseInfoKHR

VkWin32KeyedMutexAcquireReleaseInfoKHR - Use the Windows keyed mutex mechanism to synchronize work

Valid Usage

Each member of pAcquireSyncs and pReleaseSyncs must be a device memory object imported by setting ImportMemoryWin32HandleInfoKHR::handleType to EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_BIT or EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_KMT_BIT

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_WIN32_KEYED_MUTEX_ACQUIRE_RELEASE_INFO_KHR

If acquireCount is not 0, pAcquireSyncs must be a valid pointer to an array of acquireCount valid DeviceMemory handles

If acquireCount is not 0, pAcquireKeys must be a valid pointer to an array of acquireCount uint64_t values

If acquireCount is not 0, pAcquireTimeouts must be a valid pointer to an array of acquireCount uint32_t values

If releaseCount is not 0, pReleaseSyncs must be a valid pointer to an array of releaseCount valid DeviceMemory handles

If releaseCount is not 0, pReleaseKeys must be a valid pointer to an array of releaseCount uint64_t values

Both of the elements of pAcquireSyncs, and the elements of pReleaseSyncs that are valid handles of non-ignored parameters must have been created, allocated, or retrieved from the same Device

See Also

DeviceMemory, StructureType

Constructors

Win32KeyedMutexAcquireReleaseInfoKHR

Fields

acquireSyncs :: Vector DeviceMemory

pAcquireSyncs is a pointer to an array of DeviceMemory objects which were imported from Direct3D 11 resources.

acquireKeys :: Vector Word64

pAcquireKeys is a pointer to an array of mutex key values to wait for prior to beginning the submitted work. Entries refer to the keyed mutex associated with the corresponding entries in pAcquireSyncs.

acquireTimeouts :: Vector Word32

releaseSyncs :: Vector DeviceMemory

pReleaseSyncs is a pointer to an array of DeviceMemory objects which were imported from Direct3D 11 resources.

releaseKeys :: Vector Word64

pReleaseKeys is a pointer to an array of mutex key values to set when the submitted work has completed. Entries refer to the keyed mutex associated with the corresponding entries in pReleaseSyncs.

type KHR_WIN32_KEYED_MUTEX_SPEC_VERSION = 1

pattern KHR_WIN32_KEYED_MUTEX_SPEC_VERSION :: forall a. Integral a => a

type KHR_WIN32_KEYED_MUTEX_EXTENSION_NAME = "VK_KHR_win32_keyed_mutex"

pattern KHR_WIN32_KEYED_MUTEX_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a