Vulkan.Extensions.VK_MVK_macos_surface

Description

Name

VK_MVK_macos_surface - instance extension

VK_MVK_macos_surface

Name String

VK_MVK_macos_surface

Extension Type

Instance extension

Registered Extension Number

124

Revision

3

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_surface

Deprecation state

Deprecated by VK_EXT_metal_surface extension

Contact

Bill Hollings https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_MVK_macos_surface:%20&body=@billhollings%20

Other Extension Metadata

Last Modified Date

2020-07-31

IP Status

No known IP claims.

Contributors

Bill Hollings, The Brenwill Workshop Ltd.

Description

The VK_MVK_macos_surface extension is an instance extension. It provides a mechanism to create a SurfaceKHR object (defined by the VK_KHR_surface extension) based on an NSView, the native surface type of macOS, which is underpinned by a CAMetalLayer, to support rendering to the surface using Apple’s Metal framework.

Deprecation by VK_EXT_metal_surface

The VK_MVK_macos_surface extension is considered deprecated and has been superseded by the VK_EXT_metal_surface extension.

New Commands

createMacOSSurfaceMVK

New Structures

MacOSSurfaceCreateInfoMVK

New Bitmasks

MacOSSurfaceCreateFlagsMVK

New Enum Constants

MVK_MACOS_SURFACE_EXTENSION_NAME

MVK_MACOS_SURFACE_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_MACOS_SURFACE_CREATE_INFO_MVK

Version History

Revision 1, 2017-02-15 (Bill Hollings)

Initial draft.

Revision 2, 2017-02-24 (Bill Hollings)

Minor syntax fix to emphasize firm requirement for NSView to be backed by a CAMetalLayer.

Revision 3, 2020-07-31 (Bill Hollings)

Update documentation on requirements for NSView.

Mark as deprecated by VK_EXT_metal_surface.

See Also

MacOSSurfaceCreateFlagsMVK, MacOSSurfaceCreateInfoMVK, createMacOSSurfaceMVK

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

createMacOSSurfaceMVK :: forall io. MonadIO io => Instance -> MacOSSurfaceCreateInfoMVK -> ("allocator" ::: Maybe AllocationCallbacks) -> io SurfaceKHR

vkCreateMacOSSurfaceMVK - Create a VkSurfaceKHR object for a macOS NSView

Parameters

Note

The createMacOSSurfaceMVK function is considered deprecated and has been superseded by createMetalSurfaceEXT from the VK_EXT_metal_surface extension.

Description

instance is the instance with which to associate the surface.

pCreateInfo is a pointer to a MacOSSurfaceCreateInfoMVK structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface is a pointer to a SurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

instance must be a valid Instance handle

pCreateInfo must be a valid pointer to a valid MacOSSurfaceCreateInfoMVK structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pSurface must be a valid pointer to a SurfaceKHR handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_NATIVE_WINDOW_IN_USE_KHR

See Also

AllocationCallbacks, Instance, MacOSSurfaceCreateInfoMVK, SurfaceKHR

data MacOSSurfaceCreateInfoMVK

VkMacOSSurfaceCreateInfoMVK - Structure specifying parameters of a newly created macOS surface object

Valid Usage

If pView is a CAMetalLayer object, it must be a valid CAMetalLayer.

If pView is an NSView object, it must be a valid NSView, must be backed by a CALayer object of type CAMetalLayer, and createMacOSSurfaceMVK must be called on the main thread.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_MACOS_SURFACE_CREATE_INFO_MVK

pNext must be NULL

flags must be 0

See Also

MacOSSurfaceCreateFlagsMVK, StructureType, createMacOSSurfaceMVK

Constructors

MacOSSurfaceCreateInfoMVK

Fields

flags :: MacOSSurfaceCreateFlagsMVK

flags is reserved for future use.

view :: Ptr ()

pView is a reference to either a CAMetalLayer object or an NSView object.

newtype MacOSSurfaceCreateFlagsMVK

VkMacOSSurfaceCreateFlagsMVK - Reserved for future use

Description

MacOSSurfaceCreateFlagsMVK is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

MacOSSurfaceCreateInfoMVK

Constructors

MacOSSurfaceCreateFlagsMVK Flags

type MVK_MACOS_SURFACE_SPEC_VERSION = 3

pattern MVK_MACOS_SURFACE_SPEC_VERSION :: forall a. Integral a => a

type MVK_MACOS_SURFACE_EXTENSION_NAME = "VK_MVK_macos_surface"

pattern MVK_MACOS_SURFACE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype SurfaceKHR

VkSurfaceKHR - Opaque handle to a surface object

Description

The VK_KHR_surface extension declares the SurfaceKHR object, and provides a function for destroying SurfaceKHR objects. Separate platform-specific extensions each provide a function for creating a SurfaceKHR object for the respective platform. From the application’s perspective this is an opaque handle, just like the handles of other Vulkan objects.

See Also

PhysicalDeviceSurfaceInfo2KHR, SwapchainCreateInfoKHR, createAndroidSurfaceKHR, createDirectFBSurfaceEXT, createDisplayPlaneSurfaceKHR, createHeadlessSurfaceEXT, createIOSSurfaceMVK, createImagePipeSurfaceFUCHSIA, createMacOSSurfaceMVK, createMetalSurfaceEXT, createStreamDescriptorSurfaceGGP, createViSurfaceNN, createWaylandSurfaceKHR, createWin32SurfaceKHR, createXcbSurfaceKHR, createXlibSurfaceKHR, destroySurfaceKHR, getDeviceGroupSurfacePresentModesKHR, getPhysicalDevicePresentRectanglesKHR, getPhysicalDeviceSurfaceCapabilities2EXT, getPhysicalDeviceSurfaceCapabilitiesKHR, getPhysicalDeviceSurfaceFormatsKHR, getPhysicalDeviceSurfacePresentModesKHR, getPhysicalDeviceSurfaceSupportKHR

Constructors

SurfaceKHR Word64