Vulkan.Extensions.VK_NVX_image_view_handle

Documentation

getImageViewHandleNVX

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that owns the image view.

device must be a valid Device handle

-> ImageViewHandleInfoNVX
pInfo describes the image view to query and type of handle.

pInfo must be a valid pointer to a valid ImageViewHandleInfoNVX structure

-> io Word32

vkGetImageViewHandleNVX - Get the handle for an image view for a specific descriptor type

Valid Usage (Implicit)

See Also

Device, ImageViewHandleInfoNVX

getImageViewAddressNVX

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that owns the image view.

device must be a valid Device handle

-> ImageView
imageView is a handle to the image view.

imageView must be a valid ImageView handle

imageView must have been created, allocated, or retrieved from device

-> io ImageViewAddressPropertiesNVX

vkGetImageViewAddressNVX - Get the device address of an image view

Return Codes

Success

SUCCESS

Failure

ERROR_UNKNOWN

See Also

Device, ImageView, ImageViewAddressPropertiesNVX

data ImageViewHandleInfoNVX

VkImageViewHandleInfoNVX - Structure specifying the image view for handle queries

Valid Usage

descriptorType must be DESCRIPTOR_TYPE_SAMPLED_IMAGE, DESCRIPTOR_TYPE_STORAGE_IMAGE, or DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER

sampler must be a valid Sampler if descriptorType is DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER

If descriptorType is DESCRIPTOR_TYPE_SAMPLED_IMAGE or DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, the image that imageView was created from must have been created with the IMAGE_USAGE_SAMPLED_BIT usage bit set

If descriptorType is DESCRIPTOR_TYPE_STORAGE_IMAGE, the image that imageView was created from must have been created with the IMAGE_USAGE_STORAGE_BIT usage bit set

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_IMAGE_VIEW_HANDLE_INFO_NVX

pNext must be NULL

imageView must be a valid ImageView handle

descriptorType must be a valid DescriptorType value

If sampler is not NULL_HANDLE, sampler must be a valid Sampler handle

Both of imageView, and sampler that are valid handles of non-ignored parameters must have been created, allocated, or retrieved from the same Device

See Also

DescriptorType, ImageView, Sampler, StructureType, getImageViewHandleNVX

Constructors

ImageViewHandleInfoNVX

Fields

imageView :: ImageView

imageView is the image view to query.

descriptorType :: DescriptorType

descriptorType is the type of descriptor for which to query a handle.

sampler :: Sampler

sampler is the sampler to combine with the image view when generating the handle.

data ImageViewAddressPropertiesNVX

VkImageViewAddressPropertiesNVX - Structure specifying the image view for handle queries

Valid Usage (Implicit)

See Also

DeviceAddress, DeviceSize, StructureType, getImageViewAddressNVX

Constructors

ImageViewAddressPropertiesNVX

Fields

deviceAddress :: DeviceAddress

deviceAddress is the device address of the image view.

size :: DeviceSize

size is the size in bytes of the image view device memory.

type NVX_IMAGE_VIEW_HANDLE_SPEC_VERSION = 2

pattern NVX_IMAGE_VIEW_HANDLE_SPEC_VERSION :: forall a. Integral a => a

type NVX_IMAGE_VIEW_HANDLE_EXTENSION_NAME = "VK_NVX_image_view_handle"

pattern NVX_IMAGE_VIEW_HANDLE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a