Vulkan.Extensions.VK_NVX_multiview_per_view_attributes

Description

Name

VK_NVX_multiview_per_view_attributes - device extension

VK_NVX_multiview_per_view_attributes

Name String

VK_NVX_multiview_per_view_attributes

Extension Type

Device extension

Registered Extension Number

98

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_multiview

Contact

Jeff Bolz https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_NVX_multiview_per_view_attributes:%20&body=@jeffbolznv%20

Other Extension Metadata

Last Modified Date

2017-01-13

IP Status

No known IP claims.

Interactions and External Dependencies

This extension requires SPV_NVX_multiview_per_view_attributes

This extension provides API support for GL_NVX_multiview_per_view_attributes

This extension interacts with VK_NV_viewport_array2.

Contributors

Jeff Bolz, NVIDIA

Daniel Koch, NVIDIA

Description

This extension adds a new way to write shaders to be used with multiview subpasses, where the attributes for all views are written out by a single invocation of the vertex processing stages. Related SPIR-V and GLSL extensions SPV_NVX_multiview_per_view_attributes and GL_NVX_multiview_per_view_attributes introduce per-view position and viewport mask attributes arrays, and this extension defines how those per-view attribute arrays are interpreted by Vulkan. Pipelines using per-view attributes may only execute the vertex processing stages once for all views rather than once per-view, which reduces redundant shading work.

A subpass creation flag controls whether the subpass uses this extension. A subpass must either exclusively use this extension or not use it at all.

Some Vulkan implementations only support the position attribute varying between views in the X component. A subpass can declare via a second creation flag whether all pipelines compiled for this subpass will obey this restriction.

Shaders that use the new per-view outputs (e.g. gl_PositionPerViewNV) must also write the non-per-view output (gl_Position), and the values written must be such that gl_Position = gl_PositionPerViewNV[gl_ViewIndex] for all views in the subpass. Implementations are free to either use the per-view outputs or the non-per-view outputs, whichever would be more efficient.

If VK_NV_viewport_array2 is not also supported and enabled, the per-view viewport mask must not be used.

New Structures

Extending PhysicalDeviceProperties2:

PhysicalDeviceMultiviewPerViewAttributesPropertiesNVX

New Enum Constants

NVX_MULTIVIEW_PER_VIEW_ATTRIBUTES_EXTENSION_NAME

NVX_MULTIVIEW_PER_VIEW_ATTRIBUTES_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_PHYSICAL_DEVICE_MULTIVIEW_PER_VIEW_ATTRIBUTES_PROPERTIES_NVX

Extending SubpassDescriptionFlagBits:

SUBPASS_DESCRIPTION_PER_VIEW_ATTRIBUTES_BIT_NVX

SUBPASS_DESCRIPTION_PER_VIEW_POSITION_X_ONLY_BIT_NVX

New Built-In Variables

PositionPerViewNV

ViewportMaskPerViewNV

New SPIR-V Capabilities

PerViewAttributesNV

Examples

#version 450 core

#extension GL_KHX_multiview : enable
#extension GL_NVX_multiview_per_view_attributes : enable

layout(location = 0) in vec4 position;
layout(set = 0, binding = 0) uniform Block { mat4 mvpPerView[2]; } buf;

void main()
{
    // Output both per-view positions and gl_Position as a function
    // of gl_ViewIndex
    gl_PositionPerViewNV[0] = buf.mvpPerView[0] * position;
    gl_PositionPerViewNV[1] = buf.mvpPerView[1] * position;
    gl_Position = buf.mvpPerView[gl_ViewIndex] * position;
}

Version History

Revision 1, 2017-01-13 (Jeff Bolz)

Internal revisions

See Also

PhysicalDeviceMultiviewPerViewAttributesPropertiesNVX

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data PhysicalDeviceMultiviewPerViewAttributesPropertiesNVX

VkPhysicalDeviceMultiviewPerViewAttributesPropertiesNVX - Structure describing multiview limits that can be supported by an implementation

Members

The members of the PhysicalDeviceMultiviewPerViewAttributesPropertiesNVX structure describe the following implementation-dependent limits:

Description

If the PhysicalDeviceMultiviewPerViewAttributesPropertiesNVX structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceMultiviewPerViewAttributesPropertiesNVX

Fields

perViewPositionAllComponents :: Bool

perViewPositionAllComponents is TRUE if the implementation supports per-view position values that differ in components other than the X component.

type NVX_MULTIVIEW_PER_VIEW_ATTRIBUTES_SPEC_VERSION = 1

pattern NVX_MULTIVIEW_PER_VIEW_ATTRIBUTES_SPEC_VERSION :: forall a. Integral a => a

type NVX_MULTIVIEW_PER_VIEW_ATTRIBUTES_EXTENSION_NAME = "VK_NVX_multiview_per_view_attributes"

pattern NVX_MULTIVIEW_PER_VIEW_ATTRIBUTES_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a