Vulkan.Extensions.VK_NV_coverage_reduction_mode

Documentation

getPhysicalDeviceSupportedFramebufferMixedSamplesCombinationsNV

Arguments

:: forall io. MonadIO io

=> PhysicalDevice
physicalDevice is the physical device from which to query the set of combinations.

-> io (Result, "combinations" ::: Vector FramebufferMixedSamplesCombinationNV)

vkGetPhysicalDeviceSupportedFramebufferMixedSamplesCombinationsNV - Query supported sample count combinations

Description

If pCombinations is NULL, then the number of supported combinations for the given physicalDevice is returned in pCombinationCount. Otherwise, pCombinationCount must point to a variable set by the user to the number of elements in the pCombinations array, and on return the variable is overwritten with the number of values actually written to pCombinations. If the value of pCombinationCount is less than the number of combinations supported for the given physicalDevice, at most pCombinationCount values will be written pCombinations and INCOMPLETE will be returned instead of SUCCESS to indicate that not all the supported values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid PhysicalDevice handle

pCombinationCount must be a valid pointer to a uint32_t value

If the value referenced by pCombinationCount is not 0, and pCombinations is not NULL, pCombinations must be a valid pointer to an array of pCombinationCount FramebufferMixedSamplesCombinationNV structures

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

FramebufferMixedSamplesCombinationNV, PhysicalDevice

data PhysicalDeviceCoverageReductionModeFeaturesNV

VkPhysicalDeviceCoverageReductionModeFeaturesNV - Structure describing the coverage reduction mode features that can be supported by an implementation

Members

The members of the PhysicalDeviceCoverageReductionModeFeaturesNV structure describe the following features:

Description

If the PhysicalDeviceCoverageReductionModeFeaturesNV structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether the feature is supported. PhysicalDeviceCoverageReductionModeFeaturesNV can also be included in the pNext chain of DeviceCreateInfo to enable the feature.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceCoverageReductionModeFeaturesNV

Fields

coverageReductionMode :: Bool

coverageReductionMode indicates whether the implementation supports coverage reduction modes. See Coverage Reduction.

data PipelineCoverageReductionStateCreateInfoNV

VkPipelineCoverageReductionStateCreateInfoNV - Structure specifying parameters controlling coverage reduction

Description

If this structure is not present, or if the extension is not enabled, the default coverage reduction mode is inferred as follows:

If the VK_NV_framebuffer_mixed_samples extension is enabled, then it is as if the coverageReductionMode is COVERAGE_REDUCTION_MODE_MERGE_NV.

If the VK_AMD_mixed_attachment_samples extension is enabled, then it is as if the coverageReductionMode is COVERAGE_REDUCTION_MODE_TRUNCATE_NV.

If both VK_NV_framebuffer_mixed_samples and VK_AMD_mixed_attachment_samples are enabled, then the default coverage reduction mode is implementation-dependent.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PIPELINE_COVERAGE_REDUCTION_STATE_CREATE_INFO_NV

flags must be 0

coverageReductionMode must be a valid CoverageReductionModeNV value

See Also

CoverageReductionModeNV, PipelineCoverageReductionStateCreateFlagsNV, StructureType

Constructors

PipelineCoverageReductionStateCreateInfoNV

Fields

flags :: PipelineCoverageReductionStateCreateFlagsNV

flags is reserved for future use.

coverageReductionMode :: CoverageReductionModeNV

coverageReductionMode is a CoverageReductionModeNV value controlling how color sample coverage is generated from pixel coverage.

data FramebufferMixedSamplesCombinationNV

VkFramebufferMixedSamplesCombinationNV - Structure specifying a supported sample count combination

Valid Usage (Implicit)

See Also

CoverageReductionModeNV, SampleCountFlagBits, SampleCountFlags, StructureType, getPhysicalDeviceSupportedFramebufferMixedSamplesCombinationsNV

Constructors

FramebufferMixedSamplesCombinationNV

Fields

coverageReductionMode :: CoverageReductionModeNV

coverageReductionMode is a CoverageReductionModeNV value specifying the coverage reduction mode.

rasterizationSamples :: SampleCountFlagBits

rasterizationSamples specifies the number of rasterization samples in the supported combination.

depthStencilSamples :: SampleCountFlags

depthStencilSamples specifies the number of samples in the depth stencil attachment in the supported combination. A value of 0 indicates the combination does not have a depth stencil attachment.

colorSamples :: SampleCountFlags

colorSamples specifies the number of color samples in a color attachment in the supported combination. A value of 0 indicates the combination does not have a color attachment.

newtype PipelineCoverageReductionStateCreateFlagsNV

VkPipelineCoverageReductionStateCreateFlagsNV - Reserved for future use

Description

PipelineCoverageReductionStateCreateFlagsNV is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

PipelineCoverageReductionStateCreateInfoNV

Constructors

PipelineCoverageReductionStateCreateFlagsNV Flags

newtype CoverageReductionModeNV

VkCoverageReductionModeNV - Specify the coverage reduction mode

See Also

FramebufferMixedSamplesCombinationNV, PipelineCoverageReductionStateCreateInfoNV

Constructors

CoverageReductionModeNV Int32

Bundled Patterns

pattern COVERAGE_REDUCTION_MODE_MERGE_NV :: CoverageReductionModeNV
COVERAGE_REDUCTION_MODE_MERGE_NV specifies that each color sample will be associated with an implementation-dependent subset of samples in the pixel coverage. If any of those associated samples are covered, the color sample is covered.

pattern COVERAGE_REDUCTION_MODE_TRUNCATE_NV :: CoverageReductionModeNV
COVERAGE_REDUCTION_MODE_TRUNCATE_NV specifies that for color samples present in the color attachments, a color sample is covered if the pixel coverage sample with the same sample index i is covered; other pixel coverage samples are discarded.

type NV_COVERAGE_REDUCTION_MODE_SPEC_VERSION = 1

pattern NV_COVERAGE_REDUCTION_MODE_SPEC_VERSION :: forall a. Integral a => a

type NV_COVERAGE_REDUCTION_MODE_EXTENSION_NAME = "VK_NV_coverage_reduction_mode"

pattern NV_COVERAGE_REDUCTION_MODE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a