Vulkan.Extensions.VK_NV_dedicated_allocation

Description

Name

VK_NV_dedicated_allocation - device extension

VK_NV_dedicated_allocation

Name String

VK_NV_dedicated_allocation

Extension Type

Device extension

Registered Extension Number

27

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Deprecation state

Deprecated by VK_KHR_dedicated_allocation extension

Which in turn was promoted to Vulkan 1.1

Contact

Jeff Bolz https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_NV_dedicated_allocation:%20&body=@jeffbolznv%20

Other Extension Metadata

Last Modified Date

2016-05-31

IP Status

No known IP claims.

Contributors

Jeff Bolz, NVIDIA

Description

This extension allows device memory to be allocated for a particular buffer or image resource, which on some devices can significantly improve the performance of that resource. Normal device memory allocations must support memory aliasing and sparse binding, which could interfere with optimizations like framebuffer compression or efficient page table usage. This is important for render targets and very large resources, but need not (and probably should not) be used for smaller resources that can benefit from suballocation.

This extension adds a few small structures to resource creation and memory allocation: a new structure that flags whether am image/buffer will have a dedicated allocation, and a structure indicating the image or buffer that an allocation will be bound to.

New Structures

Extending BufferCreateInfo:

DedicatedAllocationBufferCreateInfoNV

Extending ImageCreateInfo:

DedicatedAllocationImageCreateInfoNV

Extending MemoryAllocateInfo:

DedicatedAllocationMemoryAllocateInfoNV

New Enum Constants

NV_DEDICATED_ALLOCATION_EXTENSION_NAME

NV_DEDICATED_ALLOCATION_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_DEDICATED_ALLOCATION_BUFFER_CREATE_INFO_NV

STRUCTURE_TYPE_DEDICATED_ALLOCATION_IMAGE_CREATE_INFO_NV

STRUCTURE_TYPE_DEDICATED_ALLOCATION_MEMORY_ALLOCATE_INFO_NV

Examples

    // Create an image with
    // VkDedicatedAllocationImageCreateInfoNV::dedicatedAllocation
    // set to VK_TRUE

    VkDedicatedAllocationImageCreateInfoNV dedicatedImageInfo =
    {
        VK_STRUCTURE_TYPE_DEDICATED_ALLOCATION_IMAGE_CREATE_INFO_NV,            // sType
        NULL,                                                                   // pNext
        VK_TRUE,                                                                // dedicatedAllocation
    };

    VkImageCreateInfo imageCreateInfo =
    {
        VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO,    // sType
        &dedicatedImageInfo                     // pNext
        // Other members set as usual
    };

    VkImage image;
    VkResult result = vkCreateImage(
        device,
        &imageCreateInfo,
        NULL,                       // pAllocator
        &image);

    VkMemoryRequirements memoryRequirements;
    vkGetImageMemoryRequirements(
        device,
        image,
        &memoryRequirements);

    // Allocate memory with VkDedicatedAllocationMemoryAllocateInfoNV::image
    // pointing to the image we are allocating the memory for

    VkDedicatedAllocationMemoryAllocateInfoNV dedicatedInfo =
    {
        VK_STRUCTURE_TYPE_DEDICATED_ALLOCATION_MEMORY_ALLOCATE_INFO_NV,             // sType
        NULL,                                                                       // pNext
        image,                                                                      // image
        VK_NULL_HANDLE,                                                             // buffer
    };

    VkMemoryAllocateInfo memoryAllocateInfo =
    {
        VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO,                 // sType
        &dedicatedInfo,                                         // pNext
        memoryRequirements.size,                                // allocationSize
        FindMemoryTypeIndex(memoryRequirements.memoryTypeBits), // memoryTypeIndex
    };

    VkDeviceMemory memory;
    vkAllocateMemory(
        device,
        &memoryAllocateInfo,
        NULL,                       // pAllocator
        &memory);

    // Bind the image to the memory

    vkBindImageMemory(
        device,
        image,
        memory,
        0);

Version History

Revision 1, 2016-05-31 (Jeff Bolz)

Internal revisions

See Also

DedicatedAllocationBufferCreateInfoNV, DedicatedAllocationImageCreateInfoNV, DedicatedAllocationMemoryAllocateInfoNV

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data DedicatedAllocationImageCreateInfoNV

VkDedicatedAllocationImageCreateInfoNV - Specify that an image is bound to a dedicated memory resource

Description

Note

Using a dedicated allocation for color and depth/stencil attachments or other large images may improve performance on some devices.

Valid Usage

If dedicatedAllocation is TRUE, ImageCreateInfo::flags must not include IMAGE_CREATE_SPARSE_BINDING_BIT, IMAGE_CREATE_SPARSE_RESIDENCY_BIT, or IMAGE_CREATE_SPARSE_ALIASED_BIT

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_DEDICATED_ALLOCATION_IMAGE_CREATE_INFO_NV

See Also

Bool32, StructureType

Constructors

DedicatedAllocationImageCreateInfoNV

Fields

dedicatedAllocation :: Bool

dedicatedAllocation specifies whether the image will have a dedicated allocation bound to it.

data DedicatedAllocationBufferCreateInfoNV

VkDedicatedAllocationBufferCreateInfoNV - Specify that a buffer is bound to a dedicated memory resource

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

DedicatedAllocationBufferCreateInfoNV

Fields

dedicatedAllocation :: Bool

dedicatedAllocation specifies whether the buffer will have a dedicated allocation bound to it.

data DedicatedAllocationMemoryAllocateInfoNV

VkDedicatedAllocationMemoryAllocateInfoNV - Specify a dedicated memory allocation resource

Valid Usage

At least one of image and buffer must be NULL_HANDLE

If image is not NULL_HANDLE, the image must have been created with DedicatedAllocationImageCreateInfoNV::dedicatedAllocation equal to TRUE

If buffer is not NULL_HANDLE, the buffer must have been created with DedicatedAllocationBufferCreateInfoNV::dedicatedAllocation equal to TRUE

If image is not NULL_HANDLE, MemoryAllocateInfo::allocationSize must equal the MemoryRequirements::size of the image

If buffer is not NULL_HANDLE, MemoryAllocateInfo::allocationSize must equal the MemoryRequirements::size of the buffer

If image is not NULL_HANDLE and MemoryAllocateInfo defines a memory import operation, the memory being imported must also be a dedicated image allocation and image must be identical to the image associated with the imported memory

If buffer is not NULL_HANDLE and MemoryAllocateInfo defines a memory import operation, the memory being imported must also be a dedicated buffer allocation and buffer must be identical to the buffer associated with the imported memory

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_DEDICATED_ALLOCATION_MEMORY_ALLOCATE_INFO_NV

If image is not NULL_HANDLE, image must be a valid Image handle

If buffer is not NULL_HANDLE, buffer must be a valid Buffer handle

Both of buffer, and image that are valid handles of non-ignored parameters must have been created, allocated, or retrieved from the same Device

See Also

Buffer, Image, StructureType

Constructors

DedicatedAllocationMemoryAllocateInfoNV

Fields

image :: Image

image is NULL_HANDLE or a handle of an image which this memory will be bound to.

buffer :: Buffer

buffer is NULL_HANDLE or a handle of a buffer which this memory will be bound to.

type NV_DEDICATED_ALLOCATION_SPEC_VERSION = 1

pattern NV_DEDICATED_ALLOCATION_SPEC_VERSION :: forall a. Integral a => a

type NV_DEDICATED_ALLOCATION_EXTENSION_NAME = "VK_NV_dedicated_allocation"

pattern NV_DEDICATED_ALLOCATION_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a