Vulkan.Extensions.VK_NV_dedicated_allocation_image_aliasing

Description

Name

VK_NV_dedicated_allocation_image_aliasing - device extension

VK_NV_dedicated_allocation_image_aliasing

Name String

VK_NV_dedicated_allocation_image_aliasing

Extension Type

Device extension

Registered Extension Number

241

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_dedicated_allocation

Contact

Nuno Subtil https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_NV_dedicated_allocation_image_aliasing:%20&body=@nsubtil%20

Other Extension Metadata

Last Modified Date

2019-01-04

Contributors

Nuno Subtil, NVIDIA

Jeff Bolz, NVIDIA

Eric Werness, NVIDIA

Axel Gneiting, id Software

Description

This extension allows applications to alias images on dedicated allocations, subject to specific restrictions: the extent and the number of layers in the image being aliased must be smaller than or equal to those of the original image for which the allocation was created, and every other image parameter must match.

New Structures

Extending PhysicalDeviceFeatures2, DeviceCreateInfo:

PhysicalDeviceDedicatedAllocationImageAliasingFeaturesNV

New Enum Constants

NV_DEDICATED_ALLOCATION_IMAGE_ALIASING_EXTENSION_NAME

NV_DEDICATED_ALLOCATION_IMAGE_ALIASING_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_PHYSICAL_DEVICE_DEDICATED_ALLOCATION_IMAGE_ALIASING_FEATURES_NV

Version History

Revision 1, 2019-01-04 (Nuno Subtil)

Internal revisions

See Also

PhysicalDeviceDedicatedAllocationImageAliasingFeaturesNV

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data PhysicalDeviceDedicatedAllocationImageAliasingFeaturesNV

VkPhysicalDeviceDedicatedAllocationImageAliasingFeaturesNV - Structure describing dedicated allocation image aliasing features that can be supported by an implementation

Members

The members of the PhysicalDeviceDedicatedAllocationImageAliasingFeaturesNV structure describe the following features:

Description

If the PhysicalDeviceDedicatedAllocationImageAliasingFeaturesNV structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether each feature is supported. PhysicalDeviceDedicatedAllocationImageAliasingFeaturesNV can also be included in the pNext chain of DeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceDedicatedAllocationImageAliasingFeaturesNV

Fields

dedicatedAllocationImageAliasing :: Bool

dedicatedAllocationImageAliasing indicates that the implementation supports aliasing of compatible image objects on a dedicated allocation.

type NV_DEDICATED_ALLOCATION_IMAGE_ALIASING_SPEC_VERSION = 1

pattern NV_DEDICATED_ALLOCATION_IMAGE_ALIASING_SPEC_VERSION :: forall a. Integral a => a

type NV_DEDICATED_ALLOCATION_IMAGE_ALIASING_EXTENSION_NAME = "VK_NV_dedicated_allocation_image_aliasing"

pattern NV_DEDICATED_ALLOCATION_IMAGE_ALIASING_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a