Vulkan.Extensions.VK_NV_device_diagnostic_checkpoints

Documentation

cmdSetCheckpointNV

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer that will receive the marker

-> ("checkpointMarker" ::: Ptr ())
pCheckpointMarker is an opaque application-provided value that will be associated with the checkpoint.

-> io ()

vkCmdSetCheckpointNV - insert diagnostic checkpoint in command stream

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, compute, or transfer operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute Transfer

See Also

CommandBuffer

getQueueCheckpointDataNV

Arguments

:: forall io. MonadIO io

=> Queue
queue is the Queue object the caller would like to retrieve checkpoint data for

-> io ("checkpointData" ::: Vector CheckpointDataNV)

vkGetQueueCheckpointDataNV - retrieve diagnostic checkpoint data

Description

If pCheckpointData is NULL, then the number of checkpoint markers available is returned in pCheckpointDataCount.

Otherwise, pCheckpointDataCount must point to a variable set by the user to the number of elements in the pCheckpointData array, and on return the variable is overwritten with the number of structures actually written to pCheckpointData.

If pCheckpointDataCount is less than the number of checkpoint markers available, at most pCheckpointDataCount structures will be written.

Valid Usage

The device that queue belongs to must be in the lost state

Valid Usage (Implicit)

queue must be a valid Queue handle

pCheckpointDataCount must be a valid pointer to a uint32_t value

If the value referenced by pCheckpointDataCount is not 0, and pCheckpointData is not NULL, pCheckpointData must be a valid pointer to an array of pCheckpointDataCount CheckpointDataNV structures

See Also

CheckpointDataNV, Queue

data QueueFamilyCheckpointPropertiesNV

VkQueueFamilyCheckpointPropertiesNV - return structure for queue family checkpoint info query

Valid Usage (Implicit)

See Also

PipelineStageFlags, StructureType

Constructors

QueueFamilyCheckpointPropertiesNV

Fields

checkpointExecutionStageMask :: PipelineStageFlags

checkpointExecutionStageMask is a mask indicating which pipeline stages the implementation can execute checkpoint markers in.

data CheckpointDataNV

VkCheckpointDataNV - return structure for command buffer checkpoint data

Description

The stages at which a checkpoint marker can be executed are implementation-defined and can be queried by calling getPhysicalDeviceQueueFamilyProperties2.

Valid Usage (Implicit)

See Also

PipelineStageFlagBits, StructureType, getQueueCheckpointDataNV

Constructors

CheckpointDataNV

Fields

stage :: PipelineStageFlagBits

stage indicates which pipeline stage the checkpoint marker data refers to.

checkpointMarker :: Ptr ()

pCheckpointMarker contains the value of the last checkpoint marker executed in the stage that stage refers to.

type NV_DEVICE_DIAGNOSTIC_CHECKPOINTS_SPEC_VERSION = 2

pattern NV_DEVICE_DIAGNOSTIC_CHECKPOINTS_SPEC_VERSION :: forall a. Integral a => a

type NV_DEVICE_DIAGNOSTIC_CHECKPOINTS_EXTENSION_NAME = "VK_NV_device_diagnostic_checkpoints"

pattern NV_DEVICE_DIAGNOSTIC_CHECKPOINTS_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a