Vulkan.Extensions.VK_NV_device_generated_commands

Documentation

cmdExecuteGeneratedCommandsNV

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command is recorded.

-> ("isPreprocessed" ::: Bool)
isPreprocessed represents whether the input data has already been preprocessed on the device. If it is FALSE this command will implicitly trigger the preprocessing step, otherwise not.

-> GeneratedCommandsInfoNV
pGeneratedCommandsInfo is a pointer to an instance of the GeneratedCommandsInfoNV structure containing parameters affecting the generation of commands.

-> io ()

vkCmdExecuteGeneratedCommandsNV - Performs the generation and execution of commands on the device

Valid Usage

[[VUID-{refpage}-None-02690]] If a ImageView is sampled with FILTER_LINEAR as a result of this command, then the image view’s format features must contain FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_LINEAR_BIT

[[VUID-{refpage}-None-02691]] If a ImageView is accessed using atomic operations as a result of this command, then the image view’s format features must contain FORMAT_FEATURE_STORAGE_IMAGE_ATOMIC_BIT

[[VUID-{refpage}-None-02692]] If a ImageView is sampled with FILTER_CUBIC_EXT as a result of this command, then the image view’s format features must contain FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_CUBIC_BIT_EXT

[[VUID-{refpage}-filterCubic-02694]] Any ImageView being sampled with FILTER_CUBIC_EXT as a result of this command must have a ImageViewType and format that supports cubic filtering, as specified by FilterCubicImageViewImageFormatPropertiesEXT::filterCubic returned by getPhysicalDeviceImageFormatProperties2

[[VUID-{refpage}-filterCubicMinmax-02695]] Any ImageView being sampled with FILTER_CUBIC_EXT with a reduction mode of either SAMPLER_REDUCTION_MODE_MIN or SAMPLER_REDUCTION_MODE_MAX as a result of this command must have a ImageViewType and format that supports cubic filtering together with minmax filtering, as specified by FilterCubicImageViewImageFormatPropertiesEXT::filterCubicMinmax returned by getPhysicalDeviceImageFormatProperties2

[[VUID-{refpage}-flags-02696]] Any Image created with a ImageCreateInfo::flags containing IMAGE_CREATE_CORNER_SAMPLED_BIT_NV sampled as a result of this command must only be sampled using a SamplerAddressMode of SAMPLER_ADDRESS_MODE_CLAMP_TO_EDGE

[[VUID-{refpage}-None-02697]] For each set n that is statically used by the Pipeline bound to the pipeline bind point used by this command, a descriptor set must have been bound to n at the same pipeline bind point, with a PipelineLayout that is compatible for set n, with the PipelineLayout used to create the current Pipeline, as described in ???

[[VUID-{refpage}-None-02698]] For each push constant that is statically used by the Pipeline bound to the pipeline bind point used by this command, a push constant value must have been set for the same pipeline bind point, with a PipelineLayout that is compatible for push constants, with the PipelineLayout used to create the current Pipeline, as described in ???

[[VUID-{refpage}-None-02699]] Descriptors in each bound descriptor set, specified via cmdBindDescriptorSets, must be valid if they are statically used by the Pipeline bound to the pipeline bind point used by this command

[[VUID-{refpage}-None-02700]] A valid pipeline must be bound to the pipeline bind point used by this command

[[VUID-{refpage}-commandBuffer-02701]] If the Pipeline object bound to the pipeline bind point used by this command requires any dynamic state, that state must have been set for commandBuffer, and done so after any previously bound pipeline with the corresponding state not specified as dynamic

[[VUID-{refpage}-None-02859]] There must not have been any calls to dynamic state setting commands for any state not specified as dynamic in the Pipeline object bound to the pipeline bind point used by this command, since that pipeline was bound

[[VUID-{refpage}-None-02702]] If the Pipeline object bound to the pipeline bind point used by this command accesses a Sampler object that uses unnormalized coordinates, that sampler must not be used to sample from any Image with a ImageView of the type IMAGE_VIEW_TYPE_3D, IMAGE_VIEW_TYPE_CUBE, IMAGE_VIEW_TYPE_1D_ARRAY, IMAGE_VIEW_TYPE_2D_ARRAY or IMAGE_VIEW_TYPE_CUBE_ARRAY, in any shader stage

[[VUID-{refpage}-None-02703]] If the Pipeline object bound to the pipeline bind point used by this command accesses a Sampler object that uses unnormalized coordinates, that sampler must not be used with any of the SPIR-V OpImageSample* or OpImageSparseSample* instructions with ImplicitLod, Dref or Proj in their name, in any shader stage

[[VUID-{refpage}-None-02704]] If the Pipeline object bound to the pipeline bind point used by this command accesses a Sampler object that uses unnormalized coordinates, that sampler must not be used with any of the SPIR-V OpImageSample* or OpImageSparseSample* instructions that includes a LOD bias or any offset values, in any shader stage

[[VUID-{refpage}-None-02705]] If the robust buffer access feature is not enabled, and if the Pipeline object bound to the pipeline bind point used by this command accesses a uniform buffer, it must not access values outside of the range of the buffer as specified in the descriptor set bound to the same pipeline bind point

[[VUID-{refpage}-None-02706]] If the robust buffer access feature is not enabled, and if the Pipeline object bound to the pipeline bind point used by this command accesses a storage buffer, it must not access values outside of the range of the buffer as specified in the descriptor set bound to the same pipeline bind point

[[VUID-{refpage}-commandBuffer-02707]] If commandBuffer is an unprotected command buffer, any resource accessed by the Pipeline object bound to the pipeline bind point used by this command must not be a protected resource

[[VUID-{refpage}-renderPass-02684]] The current render pass must be compatible with the renderPass member of the GraphicsPipelineCreateInfo structure specified when creating the Pipeline bound to PIPELINE_BIND_POINT_GRAPHICS

[[VUID-{refpage}-subpass-02685]] The subpass index of the current render pass must be equal to the subpass member of the GraphicsPipelineCreateInfo structure specified when creating the Pipeline bound to PIPELINE_BIND_POINT_GRAPHICS

[[VUID-{refpage}-None-02686]] Every input attachment used by the current subpass must be bound to the pipeline via a descriptor set

[[VUID-{refpage}-None-02687]] Image subresources used as attachments in the current render pass must not be accessed in any way other than as an attachment by this command

[[VUID-{refpage}-maxMultiviewInstanceIndex-02688]] If the draw is recorded in a render pass instance with multiview enabled, the maximum instance index must be less than or equal to PhysicalDeviceMultiviewProperties::maxMultiviewInstanceIndex

[[VUID-{refpage}-sampleLocationsEnable-02689]] If the bound graphics pipeline was created with PipelineSampleLocationsStateCreateInfoEXT::sampleLocationsEnable set to TRUE and the current subpass has a depth/stencil attachment, then that attachment must have been created with the IMAGE_CREATE_SAMPLE_LOCATIONS_COMPATIBLE_DEPTH_BIT_EXT bit set

[[VUID-{refpage}-None-04007]] All vertex input bindings accessed via vertex input variables declared in the vertex shader entry point’s interface must have either valid or NULL_HANDLE buffers bound

[[VUID-{refpage}-None-04008]] If the nullDescriptor feature is not enabled, all vertex input bindings accessed via vertex input variables declared in the vertex shader entry point’s interface must not be NULL_HANDLE

[[VUID-{refpage}-None-02721]] For a given vertex buffer binding, any attribute data fetched must be entirely contained within the corresponding vertex buffer binding, as described in ???

commandBuffer must not be a protected command buffer

If isPreprocessed is TRUE then cmdPreprocessGeneratedCommandsNV must have already been executed on the device, using the same pGeneratedCommandsInfo content as well as the content of the input buffers it references (all except GeneratedCommandsInfoNV::preprocessBuffer). Furthermore pGeneratedCommandsInfo`s indirectCommandsLayout must have been created with the INDIRECT_COMMANDS_LAYOUT_USAGE_EXPLICIT_PREPROCESS_BIT_NV bit set

GeneratedCommandsInfoNV::pipeline must match the current bound pipeline at GeneratedCommandsInfoNV::pipelineBindPoint

Transform feedback must not be active

The ::deviceGeneratedCommands feature must be enabled

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pGeneratedCommandsInfo must be a valid pointer to a valid GeneratedCommandsInfoNV structure

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, or compute operations

This command must only be called inside of a render pass instance

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Inside 	 Graphics Compute

See Also

Bool32, CommandBuffer, GeneratedCommandsInfoNV

cmdPreprocessGeneratedCommandsNV

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer which does the preprocessing.

-> GeneratedCommandsInfoNV
pGeneratedCommandsInfo is a pointer to an instance of the GeneratedCommandsInfoNV structure containing parameters affecting the preprocessing step.

-> io ()

vkCmdPreprocessGeneratedCommandsNV - Performs preprocessing for generated commands

Valid Usage

commandBuffer must not be a protected command buffer

pGeneratedCommandsInfo`s indirectCommandsLayout must have been created with the INDIRECT_COMMANDS_LAYOUT_USAGE_EXPLICIT_PREPROCESS_BIT_NV bit set

The ::deviceGeneratedCommands feature must be enabled

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pGeneratedCommandsInfo must be a valid pointer to a valid GeneratedCommandsInfoNV structure

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, or compute operations

This command must only be called outside of a render pass instance

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Outside 	 Graphics Compute

See Also

CommandBuffer, GeneratedCommandsInfoNV

cmdBindPipelineShaderGroupNV

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer that the pipeline will be bound to.

-> PipelineBindPoint
pipelineBindPoint is a PipelineBindPoint value specifying to which bind point the pipeline is bound.

-> Pipeline
pipeline is the pipeline to be bound.

-> ("groupIndex" ::: Word32)
groupIndex is the shader group to be bound.

-> io ()

vkCmdBindPipelineShaderGroupNV - Bind a pipeline object

Valid Usage

groupIndex must be 0 or less than the effective GraphicsPipelineShaderGroupsCreateInfoNV::groupCount including the referenced pipelines

The pipelineBindPoint must be PIPELINE_BIND_POINT_GRAPHICS

The same restrictions as cmdBindPipeline apply as if the bound pipeline was created only with the Shader Group from the groupIndex information

The ::deviceGeneratedCommands feature must be enabled

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pipelineBindPoint must be a valid PipelineBindPoint value

pipeline must be a valid Pipeline handle

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, or compute operations

Both of commandBuffer, and pipeline must have been created, allocated, or retrieved from the same Device

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute

See Also

CommandBuffer, Pipeline, PipelineBindPoint

getGeneratedCommandsMemoryRequirementsNV

Arguments

:: forall a io. (Extendss MemoryRequirements2 a, PokeChain a, PeekChain a, MonadIO io)

=> Device
device is the logical device that owns the buffer.

-> GeneratedCommandsMemoryRequirementsInfoNV
pInfo is a pointer to an instance of the GeneratedCommandsMemoryRequirementsInfoNV structure containing parameters required for the memory requirements query.

-> io (MemoryRequirements2 a)

vkGetGeneratedCommandsMemoryRequirementsNV - Retrieve the buffer allocation requirements for generated commands

Valid Usage

The ::deviceGeneratedCommands feature must be enabled

Valid Usage (Implicit)

device must be a valid Device handle

pInfo must be a valid pointer to a valid GeneratedCommandsMemoryRequirementsInfoNV structure

pMemoryRequirements must be a valid pointer to a MemoryRequirements2 structure

See Also

Device, GeneratedCommandsMemoryRequirementsInfoNV, MemoryRequirements2

createIndirectCommandsLayoutNV

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that creates the indirect command layout.

-> IndirectCommandsLayoutCreateInfoNV
pCreateInfo is a pointer to an instance of the IndirectCommandsLayoutCreateInfoNV structure containing parameters affecting creation of the indirect command layout.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io IndirectCommandsLayoutNV

vkCreateIndirectCommandsLayoutNV - Create an indirect command layout object

Valid Usage

The ::deviceGeneratedCommands feature must be enabled

Valid Usage (Implicit)

device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid IndirectCommandsLayoutCreateInfoNV structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pIndirectCommandsLayout must be a valid pointer to a IndirectCommandsLayoutNV handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, Device, IndirectCommandsLayoutCreateInfoNV, IndirectCommandsLayoutNV

withIndirectCommandsLayoutNV :: forall io r. MonadIO io => Device -> IndirectCommandsLayoutCreateInfoNV -> Maybe AllocationCallbacks -> (io IndirectCommandsLayoutNV -> (IndirectCommandsLayoutNV -> io ()) -> r) -> r

A convenience wrapper to make a compatible pair of calls to createIndirectCommandsLayoutNV and destroyIndirectCommandsLayoutNV

To ensure that destroyIndirectCommandsLayoutNV is always called: pass bracket (or the allocate function from your favourite resource management library) as the first argument. To just extract the pair pass (,) as the first argument.

destroyIndirectCommandsLayoutNV

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that destroys the layout.

-> IndirectCommandsLayoutNV
indirectCommandsLayout is the layout to destroy.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io ()

vkDestroyIndirectCommandsLayoutNV - Destroy an indirect commands layout

Valid Usage

All submitted commands that refer to indirectCommandsLayout must have completed execution

If AllocationCallbacks were provided when indirectCommandsLayout was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when indirectCommandsLayout was created, pAllocator must be NULL

The ::deviceGeneratedCommands feature must be enabled

Valid Usage (Implicit)

device must be a valid Device handle

If indirectCommandsLayout is not NULL_HANDLE, indirectCommandsLayout must be a valid IndirectCommandsLayoutNV handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If indirectCommandsLayout is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to indirectCommandsLayout must be externally synchronized

See Also

AllocationCallbacks, Device, IndirectCommandsLayoutNV

data PhysicalDeviceDeviceGeneratedCommandsFeaturesNV

VkPhysicalDeviceDeviceGeneratedCommandsFeaturesNV - Structure describing the device-generated commands features that can be supported by an implementation

Members

The members of the PhysicalDeviceDeviceGeneratedCommandsFeaturesNV structure describe the following features:

Description

If the PhysicalDeviceDeviceGeneratedCommandsFeaturesNV structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether the feature is supported. PhysicalDeviceDeviceGeneratedCommandsFeaturesNV can also be used in the pNext chain of DeviceCreateInfo to enable the features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceDeviceGeneratedCommandsFeaturesNV

Fields

deviceGeneratedCommands :: Bool

deviceGeneratedCommands indicates whether the implementation supports functionality to generate commands on the device. See Device-Generated Commands.

data PhysicalDeviceDeviceGeneratedCommandsPropertiesNV

VkPhysicalDeviceDeviceGeneratedCommandsPropertiesNV - Structure describing push descriptor limits that can be supported by an implementation

Valid Usage (Implicit)

See Also

StructureType

Constructors

PhysicalDeviceDeviceGeneratedCommandsPropertiesNV

Fields

maxGraphicsShaderGroupCount :: Word32

maxGraphicsShaderGroupCount is the maximum number of shader groups in GraphicsPipelineShaderGroupsCreateInfoNV.

maxIndirectSequenceCount :: Word32

maxIndirectSequenceCount is the maximum number of sequences in GeneratedCommandsInfoNV and in GeneratedCommandsMemoryRequirementsInfoNV.

maxIndirectCommandsTokenCount :: Word32

maxIndirectCommandsStreamCount :: Word32

maxIndirectCommandsStreamCount is the maximum number of streams in IndirectCommandsLayoutCreateInfoNV.

maxIndirectCommandsTokenOffset :: Word32

maxIndirectCommandsTokenOffset is the maximum offset in IndirectCommandsLayoutTokenNV.

maxIndirectCommandsStreamStride :: Word32

maxIndirectCommandsStreamStride is the maximum stream stride in IndirectCommandsLayoutCreateInfoNV.

minSequencesCountBufferOffsetAlignment :: Word32

minSequencesIndexBufferOffsetAlignment :: Word32

minIndirectCommandsBufferOffsetAlignment :: Word32

minIndirectCommandsBufferOffsetAlignment is the minimum alignment for memory addresses used in IndirectCommandsStreamNV and as preprocess buffer in GeneratedCommandsInfoNV.

data GraphicsShaderGroupCreateInfoNV

VkGraphicsShaderGroupCreateInfoNV - Structure specifying override parameters for each shader group

Valid Usage

For stageCount, the same restrictions as in GraphicsPipelineCreateInfo::stageCount apply

For pStages, the same restrictions as in GraphicsPipelineCreateInfo::pStages apply

For pVertexInputState, the same restrictions as in GraphicsPipelineCreateInfo::pVertexInputState apply

For pTessellationState, the same restrictions as in GraphicsPipelineCreateInfo::pTessellationState apply

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_GRAPHICS_SHADER_GROUP_CREATE_INFO_NV

pNext must be NULL

pStages must be a valid pointer to an array of stageCount valid PipelineShaderStageCreateInfo structures

stageCount must be greater than 0

See Also

GraphicsPipelineShaderGroupsCreateInfoNV, PipelineShaderStageCreateInfo, PipelineTessellationStateCreateInfo, PipelineVertexInputStateCreateInfo, StructureType

Constructors

GraphicsShaderGroupCreateInfoNV

Fields

stages :: Vector (SomeStruct PipelineShaderStageCreateInfo)

pStages is an array of size stageCount structures of type PipelineShaderStageCreateInfo describing the set of the shader stages to be included in this shader group.

vertexInputState :: Maybe (SomeStruct PipelineVertexInputStateCreateInfo)

pVertexInputState is a pointer to an instance of the PipelineVertexInputStateCreateInfo structure.

tessellationState :: Maybe (SomeStruct PipelineTessellationStateCreateInfo)

pTessellationState is a pointer to an instance of the PipelineTessellationStateCreateInfo structure, and is ignored if the shader group does not include a tessellation control shader stage and tessellation evaluation shader stage.

data GraphicsPipelineShaderGroupsCreateInfoNV

VkGraphicsPipelineShaderGroupsCreateInfoNV - Structure specifying parameters of a newly created multi shader group pipeline

Description

When referencing shader groups by index, groups defined in the referenced pipelines are treated as if they were defined as additional entries in pGroups. They are appended in the order they appear in the pPipelines array and in the pGroups array when those pipelines were defined.

The application must maintain the lifetime of all such referenced pipelines based on the pipelines that make use of them.

Valid Usage

groupCount must be at least 1 and as maximum PhysicalDeviceDeviceGeneratedCommandsPropertiesNV::maxGraphicsShaderGroupCount

The sum of groupCount including those groups added from referenced pPipelines must also be as maximum PhysicalDeviceDeviceGeneratedCommandsPropertiesNV::maxGraphicsShaderGroupCount

The state of the first element of pGroups must match its equivalent within the parent’s GraphicsPipelineCreateInfo

Each element of pGroups must in combination with the rest of the pipeline state yield a valid state configuration

All elements of pGroups must use the same shader stage combinations unless any mesh shader stage is used, then either combination of task and mesh or just mesh shader is valid

Mesh and regular primitive shading stages cannot be mixed across pGroups

Each element of the pPipelines member of libraries must have been created with identical state to the pipeline currently created except the state that can be overriden by GraphicsShaderGroupCreateInfoNV

The ::deviceGeneratedCommands feature must be enabled

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_GRAPHICS_PIPELINE_SHADER_GROUPS_CREATE_INFO_NV

pGroups must be a valid pointer to an array of groupCount valid GraphicsShaderGroupCreateInfoNV structures

If pipelineCount is not 0, pPipelines must be a valid pointer to an array of pipelineCount valid Pipeline handles

groupCount must be greater than 0

See Also

GraphicsShaderGroupCreateInfoNV, Pipeline, StructureType

Constructors

GraphicsPipelineShaderGroupsCreateInfoNV

Fields

groups :: Vector GraphicsShaderGroupCreateInfoNV

pGroups is an array of GraphicsShaderGroupCreateInfoNV values specifying which state of the original GraphicsPipelineCreateInfo each shader group overrides.

pipelines :: Vector Pipeline

pPipelines is an array of graphics Pipeline, which are referenced within the created pipeline, including all their shader groups.

data BindShaderGroupIndirectCommandNV

VkBindShaderGroupIndirectCommandNV - Structure specifying input data for a single shader group command token

Valid Usage

The current bound graphics pipeline, as well as the pipelines it may reference, must have been created with PIPELINE_CREATE_INDIRECT_BINDABLE_BIT_NV

The index must be within range of the accessible shader groups of the current bound graphics pipeline. See cmdBindPipelineShaderGroupNV for further details

See Also

No cross-references are available

Constructors

BindShaderGroupIndirectCommandNV

Fields

groupIndex :: Word32

data BindIndexBufferIndirectCommandNV

VkBindIndexBufferIndirectCommandNV - Structure specifying input data for a single index buffer command token

Valid Usage

The buffer’s usage flag from which the address was acquired must have the BUFFER_USAGE_INDEX_BUFFER_BIT bit set

The bufferAddress must be aligned to the indexType used

Each element of the buffer from which the address was acquired and that is non-sparse must be bound completely and contiguously to a single DeviceMemory object

Valid Usage (Implicit)

indexType must be a valid IndexType value

See Also

DeviceAddress, IndexType

Constructors

BindIndexBufferIndirectCommandNV

Fields

bufferAddress :: DeviceAddress

bufferAddress specifies a physical address of the Buffer used as index buffer.

size :: Word32

size is the byte size range which is available for this operation from the provided address.

indexType :: IndexType

indexType is a IndexType value specifying how indices are treated. Instead of the Vulkan enum values, a custom uint32_t value can be mapped to an IndexType by specifying the IndirectCommandsLayoutTokenNV::pIndexTypes and IndirectCommandsLayoutTokenNV::pIndexTypeValues arrays.

data BindVertexBufferIndirectCommandNV

VkBindVertexBufferIndirectCommandNV - Structure specifying input data for a single vertex buffer command token

Valid Usage

The buffer’s usage flag from which the address was acquired must have the BUFFER_USAGE_VERTEX_BUFFER_BIT bit set

Each element of the buffer from which the address was acquired and that is non-sparse must be bound completely and contiguously to a single DeviceMemory object

See Also

DeviceAddress

Constructors

BindVertexBufferIndirectCommandNV

Fields

bufferAddress :: DeviceAddress

bufferAddress specifies a physical address of the Buffer used as vertex input binding.

size :: Word32

size is the byte size range which is available for this operation from the provided address.

stride :: Word32

stride is the byte size stride for this vertex input binding as in VertexInputBindingDescription::stride. It is only used if IndirectCommandsLayoutTokenNV::vertexDynamicStride was set, otherwise the stride is inherited from the current bound graphics pipeline.

data SetStateFlagsIndirectCommandNV

VkSetStateFlagsIndirectCommandNV - Structure specifying input data for a single state flag command token

See Also

No cross-references are available

Constructors

SetStateFlagsIndirectCommandNV

Fields

data' :: Word32

data encodes packed state that this command alters.

Bit 0: If set represents FRONT_FACE_CLOCKWISE, otherwise FRONT_FACE_COUNTER_CLOCKWISE

data IndirectCommandsStreamNV

VkIndirectCommandsStreamNV - Structure specifying input streams for generated command tokens

Valid Usage

The buffer’s usage flag must have the BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set

The offset must be aligned to PhysicalDeviceDeviceGeneratedCommandsPropertiesNV::minIndirectCommandsBufferOffsetAlignment

If buffer is non-sparse then it must be bound completely and contiguously to a single DeviceMemory object

Valid Usage (Implicit)

buffer must be a valid Buffer handle

See Also

Buffer, DeviceSize, GeneratedCommandsInfoNV

Constructors

IndirectCommandsStreamNV

Fields

buffer :: Buffer

buffer specifies the Buffer storing the functional arguments for each sequence. These arguments can be written by the device.

offset :: DeviceSize

offset specified an offset into buffer where the arguments start.

data IndirectCommandsLayoutTokenNV

VkIndirectCommandsLayoutTokenNV - Struct specifying the details of an indirect command layout token

Valid Usage

stream must be smaller than IndirectCommandsLayoutCreateInfoNV::streamCount

offset must be less than or equal to PhysicalDeviceDeviceGeneratedCommandsPropertiesNV::maxIndirectCommandsTokenOffset

If tokenType is INDIRECT_COMMANDS_TOKEN_TYPE_VERTEX_BUFFER_NV, vertexBindingUnit must stay within device supported limits for the appropriate commands

If tokenType is INDIRECT_COMMANDS_TOKEN_TYPE_PUSH_CONSTANT_NV, pushconstantPipelineLayout must be valid

If tokenType is INDIRECT_COMMANDS_TOKEN_TYPE_PUSH_CONSTANT_NV, pushconstantOffset must be a multiple of 4

If tokenType is INDIRECT_COMMANDS_TOKEN_TYPE_PUSH_CONSTANT_NV, pushconstantSize must be a multiple of 4

If tokenType is INDIRECT_COMMANDS_TOKEN_TYPE_PUSH_CONSTANT_NV, pushconstantOffset must be less than PhysicalDeviceLimits::maxPushConstantsSize

If tokenType is INDIRECT_COMMANDS_TOKEN_TYPE_PUSH_CONSTANT_NV, pushconstantSize must be less than or equal to PhysicalDeviceLimits::maxPushConstantsSize minus pushconstantOffset

If tokenType is INDIRECT_COMMANDS_TOKEN_TYPE_PUSH_CONSTANT_NV, for each byte in the range specified by pushconstantOffset and pushconstantSize and for each shader stage in pushconstantShaderStageFlags, there must be a push constant range in pushconstantPipelineLayout that includes that byte and that stage

If tokenType is INDIRECT_COMMANDS_TOKEN_TYPE_PUSH_CONSTANT_NV, for each byte in the range specified by pushconstantOffset and pushconstantSize and for each push constant range that overlaps that byte, pushconstantShaderStageFlags must include all stages in that push constant range’s PushConstantRange::pushconstantShaderStageFlags

If tokenType is INDIRECT_COMMANDS_TOKEN_TYPE_STATE_FLAGS_NV, indirectStateFlags must not be ´0´

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_INDIRECT_COMMANDS_LAYOUT_TOKEN_NV

pNext must be NULL

tokenType must be a valid IndirectCommandsTokenTypeNV value

If pushconstantPipelineLayout is not NULL_HANDLE, pushconstantPipelineLayout must be a valid PipelineLayout handle

pushconstantShaderStageFlags must be a valid combination of ShaderStageFlagBits values

indirectStateFlags must be a valid combination of IndirectStateFlagBitsNV values

If indexTypeCount is not 0, pIndexTypes must be a valid pointer to an array of indexTypeCount valid IndexType values

If indexTypeCount is not 0, pIndexTypeValues must be a valid pointer to an array of indexTypeCount uint32_t values

See Also

Bool32, IndexType, IndirectCommandsLayoutCreateInfoNV, IndirectCommandsTokenTypeNV, IndirectStateFlagsNV, PipelineLayout, ShaderStageFlags, StructureType

Constructors

IndirectCommandsLayoutTokenNV

Fields

tokenType :: IndirectCommandsTokenTypeNV

tokenType specifies the token command type.

stream :: Word32

stream is the index of the input stream that contains the token argument data.

offset :: Word32

offset is a relative starting offset within the input stream memory for the token argument data.

vertexBindingUnit :: Word32

vertexBindingUnit is used for the vertex buffer binding command.

vertexDynamicStride :: Bool

vertexDynamicStride sets if the vertex buffer stride is provided by the binding command rather than the current bound graphics pipeline state.

pushconstantPipelineLayout :: PipelineLayout

pushconstantPipelineLayout is the PipelineLayout used for the push constant command.

pushconstantShaderStageFlags :: ShaderStageFlags

pushconstantShaderStageFlags are the shader stage flags used for the push constant command.

pushconstantOffset :: Word32

pushconstantOffset is the offset used for the push constant command.

pushconstantSize :: Word32

pushconstantSize is the size used for the push constant command.

indirectStateFlags :: IndirectStateFlagsNV

indirectStateFlags are the active states for the state flag command.

indexTypes :: Vector IndexType

pIndexTypes is the used IndexType for the corresponding uint32_t value entry in pIndexTypeValues.

indexTypeValues :: Vector Word32

data IndirectCommandsLayoutCreateInfoNV

VkIndirectCommandsLayoutCreateInfoNV - Structure specifying the parameters of a newly created indirect commands layout object

Description

The following code illustrates some of the flags:

void cmdProcessAllSequences(cmd, pipeline, indirectCommandsLayout, pIndirectCommandsTokens, sequencesCount, indexbuffer, indexbufferOffset)
{
  for (s = 0; s < sequencesCount; s++)
  {
    sUsed = s;

    if (indirectCommandsLayout.flags & VK_INDIRECT_COMMANDS_LAYOUT_USAGE_INDEXED_SEQUENCES_BIT_NV) {
      sUsed = indexbuffer.load_uint32( sUsed * sizeof(uint32_t) + indexbufferOffset);
    }

    if (indirectCommandsLayout.flags & VK_INDIRECT_COMMANDS_LAYOUT_USAGE_UNORDERED_SEQUENCES_BIT_NV) {
      sUsed = incoherent_implementation_dependent_permutation[ sUsed ];
    }

    cmdProcessSequence( cmd, pipeline, indirectCommandsLayout, pIndirectCommandsTokens, sUsed );
  }
}

Valid Usage

The pipelineBindPoint must be PIPELINE_BIND_POINT_GRAPHICS

tokenCount must be greater than 0 and less than or equal to PhysicalDeviceDeviceGeneratedCommandsPropertiesNV::maxIndirectCommandsTokenCount

If pTokens contains an entry of INDIRECT_COMMANDS_TOKEN_TYPE_SHADER_GROUP_NV it must be the first element of the array and there must be only a single element of such token type

If pTokens contains an entry of INDIRECT_COMMANDS_TOKEN_TYPE_STATE_FLAGS_NV there must be only a single element of such token type

All state tokens in pTokens must occur prior work provoking tokens (INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_NV, INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_INDEXED_NV, INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_TASKS_NV)

The content of pTokens must include one single work provoking token that is compatible with the pipelineBindPoint

streamCount must be greater than 0 and less or equal to PhysicalDeviceDeviceGeneratedCommandsPropertiesNV::maxIndirectCommandsStreamCount

each element of pStreamStrides must be greater than `0`and less than or equal to PhysicalDeviceDeviceGeneratedCommandsPropertiesNV::maxIndirectCommandsStreamStride. Furthermore the alignment of each token input must be ensured

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_INDIRECT_COMMANDS_LAYOUT_CREATE_INFO_NV

pNext must be NULL

flags must be a valid combination of IndirectCommandsLayoutUsageFlagBitsNV values

flags must not be 0

pipelineBindPoint must be a valid PipelineBindPoint value

pTokens must be a valid pointer to an array of tokenCount valid IndirectCommandsLayoutTokenNV structures

pStreamStrides must be a valid pointer to an array of streamCount uint32_t values

tokenCount must be greater than 0

streamCount must be greater than 0

See Also

IndirectCommandsLayoutTokenNV, IndirectCommandsLayoutUsageFlagsNV, PipelineBindPoint, StructureType, createIndirectCommandsLayoutNV

Constructors

IndirectCommandsLayoutCreateInfoNV

Fields

flags :: IndirectCommandsLayoutUsageFlagsNV

flags is a bitmask of IndirectCommandsLayoutUsageFlagBitsNV specifying usage hints of this layout.

pipelineBindPoint :: PipelineBindPoint

pipelineBindPoint is the PipelineBindPoint that this layout targets.

tokens :: Vector IndirectCommandsLayoutTokenNV

pTokens is an array describing each command token in detail. See IndirectCommandsTokenTypeNV and IndirectCommandsLayoutTokenNV below for details.

streamStrides :: Vector Word32

pStreamStrides is an array defining the byte stride for each input stream.

data GeneratedCommandsInfoNV

VkGeneratedCommandsInfoNV - Structure specifying parameters for the generation of commands

Valid Usage

The provided pipeline must match the pipeline bound at execution time

If the indirectCommandsLayout uses a token of INDIRECT_COMMANDS_TOKEN_TYPE_SHADER_GROUP_NV, then the pipeline must have been created with multiple shader groups

If the indirectCommandsLayout uses a token of INDIRECT_COMMANDS_TOKEN_TYPE_SHADER_GROUP_NV, then the pipeline must have been created with PIPELINE_CREATE_INDIRECT_BINDABLE_BIT_NV set in GraphicsPipelineCreateInfo::flags

If the indirectCommandsLayout uses a token of INDIRECT_COMMANDS_TOKEN_TYPE_PUSH_CONSTANT_NV, then the pipeline`s PipelineLayout must match the IndirectCommandsLayoutTokenNV::pushconstantPipelineLayout

streamCount must match the indirectCommandsLayout’s streamCount

sequencesCount must be less or equal to PhysicalDeviceDeviceGeneratedCommandsPropertiesNV::maxIndirectSequenceCount and GeneratedCommandsMemoryRequirementsInfoNV::maxSequencesCount that was used to determine the preprocessSize

preprocessBuffer must have the BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set in its usage flag

preprocessOffset must be aligned to PhysicalDeviceDeviceGeneratedCommandsPropertiesNV::minIndirectCommandsBufferOffsetAlignment

If preprocessBuffer is non-sparse then it must be bound completely and contiguously to a single DeviceMemory object

preprocessSize must be at least equal to the memory requirement`s size returned by getGeneratedCommandsMemoryRequirementsNV using the matching inputs (indirectCommandsLayout, …​) as within this structure

sequencesCountBuffer can be set if the actual used count of sequences is sourced from the provided buffer. In that case the sequencesCount serves as upper bound

If sequencesCountBuffer is not NULL_HANDLE, its usage flag must have the BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set

If sequencesCountBuffer is not NULL_HANDLE, sequencesCountOffset must be aligned to PhysicalDeviceDeviceGeneratedCommandsPropertiesNV::minSequencesCountBufferOffsetAlignment

If sequencesCountBuffer is not NULL_HANDLE and is non-sparse then it must be bound completely and contiguously to a single DeviceMemory object

If indirectCommandsLayout’s INDIRECT_COMMANDS_LAYOUT_USAGE_INDEXED_SEQUENCES_BIT_NV is set, sequencesIndexBuffer must be set otherwise it must be NULL_HANDLE

If sequencesIndexBuffer is not NULL_HANDLE, its usage flag must have the BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set

If sequencesIndexBuffer is not NULL_HANDLE, sequencesIndexOffset must be aligned to PhysicalDeviceDeviceGeneratedCommandsPropertiesNV::minSequencesIndexBufferOffsetAlignment

If sequencesIndexBuffer is not NULL_HANDLE and is non-sparse then it must be bound completely and contiguously to a single DeviceMemory object

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_GENERATED_COMMANDS_INFO_NV

pNext must be NULL

pipelineBindPoint must be a valid PipelineBindPoint value

pipeline must be a valid Pipeline handle

indirectCommandsLayout must be a valid IndirectCommandsLayoutNV handle

pStreams must be a valid pointer to an array of streamCount valid IndirectCommandsStreamNV structures

preprocessBuffer must be a valid Buffer handle

If sequencesCountBuffer is not NULL_HANDLE, sequencesCountBuffer must be a valid Buffer handle

If sequencesIndexBuffer is not NULL_HANDLE, sequencesIndexBuffer must be a valid Buffer handle

streamCount must be greater than 0

Each of indirectCommandsLayout, pipeline, preprocessBuffer, sequencesCountBuffer, and sequencesIndexBuffer that are valid handles of non-ignored parameters must have been created, allocated, or retrieved from the same Device

See Also

Buffer, DeviceSize, IndirectCommandsLayoutNV, IndirectCommandsStreamNV, Pipeline, PipelineBindPoint, StructureType, cmdExecuteGeneratedCommandsNV, cmdPreprocessGeneratedCommandsNV

Constructors

GeneratedCommandsInfoNV

Fields

pipelineBindPoint :: PipelineBindPoint

pipelineBindPoint is the PipelineBindPoint used for the pipeline.

pipeline :: Pipeline

pipeline is the Pipeline used in the generation and execution process.

indirectCommandsLayout :: IndirectCommandsLayoutNV

indirectCommandsLayout is the IndirectCommandsLayoutNV that provides the command sequence to generate.

streams :: Vector IndirectCommandsStreamNV

pStreams provides an array of IndirectCommandsStreamNV that provide the input data for the tokens used in indirectCommandsLayout.

sequencesCount :: Word32

sequencesCount is the maximum number of sequences to reserve. If sequencesCountBuffer is NULL_HANDLE, this is also the actual number of sequences generated.

preprocessBuffer :: Buffer

preprocessBuffer is the Buffer that is used for preprocessing the input data for execution. If this structure is used with cmdExecuteGeneratedCommandsNV with its isPreprocessed set to TRUE, then the preprocessing step is skipped and data is only read from this buffer.

preprocessOffset :: DeviceSize

preprocessOffset is the byte offset into preprocessBuffer where the preprocessed data is stored.

preprocessSize :: DeviceSize

preprocessSize is the maximum byte size within the preprocessBuffer after the preprocessOffset that is available for preprocessing.

sequencesCountBuffer :: Buffer

sequencesCountBuffer is a Buffer in which the actual number of sequences is provided as single uint32_t value.

sequencesCountOffset :: DeviceSize

sequencesCountOffset is the byte offset into sequencesCountBuffer where the count value is stored.

sequencesIndexBuffer :: Buffer

sequencesIndexBuffer is a Buffer that encodes the used sequence indices as uint32_t array.

sequencesIndexOffset :: DeviceSize

sequencesIndexOffset is the byte offset into sequencesIndexBuffer where the index values start.

data GeneratedCommandsMemoryRequirementsInfoNV

VkGeneratedCommandsMemoryRequirementsInfoNV - Structure specifying parameters for the reservation of preprocess buffer space

Valid Usage

maxSequencesCount must be less or equal to PhysicalDeviceDeviceGeneratedCommandsPropertiesNV::maxIndirectSequenceCount

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_GENERATED_COMMANDS_MEMORY_REQUIREMENTS_INFO_NV

pNext must be NULL

pipelineBindPoint must be a valid PipelineBindPoint value

pipeline must be a valid Pipeline handle

indirectCommandsLayout must be a valid IndirectCommandsLayoutNV handle

Both of indirectCommandsLayout, and pipeline must have been created, allocated, or retrieved from the same Device

See Also

IndirectCommandsLayoutNV, Pipeline, PipelineBindPoint, StructureType, getGeneratedCommandsMemoryRequirementsNV

Constructors

GeneratedCommandsMemoryRequirementsInfoNV

Fields

pipelineBindPoint :: PipelineBindPoint

pipelineBindPoint is the PipelineBindPoint of the pipeline that this buffer memory is intended to be used with during the execution.

pipeline :: Pipeline

pipeline is the Pipeline that this buffer memory is intended to be used with during the execution.

indirectCommandsLayout :: IndirectCommandsLayoutNV

indirectCommandsLayout is the IndirectCommandsLayoutNV that this buffer memory is intended to be used with.

maxSequencesCount :: Word32

maxSequencesCount is the maximum number of sequences that this buffer memory in combination with the other state provided can be used with.

newtype IndirectCommandsLayoutUsageFlagBitsNV

VkIndirectCommandsLayoutUsageFlagBitsNV - Bitmask specifying allowed usage of an indirect commands layout

See Also

IndirectCommandsLayoutUsageFlagsNV

Constructors

IndirectCommandsLayoutUsageFlagBitsNV Flags

Bundled Patterns

pattern INDIRECT_COMMANDS_LAYOUT_USAGE_EXPLICIT_PREPROCESS_BIT_NV :: IndirectCommandsLayoutUsageFlagBitsNV
INDIRECT_COMMANDS_LAYOUT_USAGE_EXPLICIT_PREPROCESS_BIT_NV specifies that the layout is always used with the manual preprocessing step through calling cmdPreprocessGeneratedCommandsNV and executed by cmdExecuteGeneratedCommandsNV with isPreprocessed set to TRUE.

pattern INDIRECT_COMMANDS_LAYOUT_USAGE_INDEXED_SEQUENCES_BIT_NV :: IndirectCommandsLayoutUsageFlagBitsNV
INDIRECT_COMMANDS_LAYOUT_USAGE_INDEXED_SEQUENCES_BIT_NV specifies that the input data for the sequences is not implicitly indexed from 0..sequencesUsed but a user provided Buffer encoding the index is provided.

pattern INDIRECT_COMMANDS_LAYOUT_USAGE_UNORDERED_SEQUENCES_BIT_NV :: IndirectCommandsLayoutUsageFlagBitsNV
INDIRECT_COMMANDS_LAYOUT_USAGE_UNORDERED_SEQUENCES_BIT_NV specifies that the processing of sequences can happen at an implementation-dependent order, which is not: guaranteed to be coherent using the same input data.

type IndirectCommandsLayoutUsageFlagsNV = IndirectCommandsLayoutUsageFlagBitsNV

newtype IndirectStateFlagBitsNV

VkIndirectStateFlagBitsNV - Bitmask specifiying state that can be altered on the device

See Also

IndirectStateFlagsNV

Constructors

IndirectStateFlagBitsNV Flags

Bundled Patterns

pattern INDIRECT_STATE_FLAG_FRONTFACE_BIT_NV :: IndirectStateFlagBitsNV
INDIRECT_STATE_FLAG_FRONTFACE_BIT_NV allows to toggle the FrontFace rasterization state for subsequent draw operations.

type IndirectStateFlagsNV = IndirectStateFlagBitsNV

newtype IndirectCommandsTokenTypeNV

VkIndirectCommandsTokenTypeNV - Enum specifying token commands

Description

'

Token type 	 Equivalent command

INDIRECT_COMMANDS_TOKEN_TYPE_SHADER_GROUP_NV 	 cmdBindPipelineShaderGroupNV

INDIRECT_COMMANDS_TOKEN_TYPE_STATE_FLAGS_NV 	 -

INDIRECT_COMMANDS_TOKEN_TYPE_INDEX_BUFFER_NV 	 cmdBindIndexBuffer

INDIRECT_COMMANDS_TOKEN_TYPE_VERTEX_BUFFER_NV 	 cmdBindVertexBuffers

INDIRECT_COMMANDS_TOKEN_TYPE_PUSH_CONSTANT_NV 	 cmdPushConstants

INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_INDEXED_NV 	 cmdDrawIndexedIndirect

INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_NV 	 cmdDrawIndirect

INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_TASKS_NV 	 cmdDrawMeshTasksIndirectNV

Supported indirect command tokens

See Also

IndirectCommandsLayoutTokenNV

Constructors

IndirectCommandsTokenTypeNV Int32

Bundled Patterns

pattern INDIRECT_COMMANDS_TOKEN_TYPE_SHADER_GROUP_NV :: IndirectCommandsTokenTypeNV

pattern INDIRECT_COMMANDS_TOKEN_TYPE_STATE_FLAGS_NV :: IndirectCommandsTokenTypeNV

pattern INDIRECT_COMMANDS_TOKEN_TYPE_INDEX_BUFFER_NV :: IndirectCommandsTokenTypeNV

pattern INDIRECT_COMMANDS_TOKEN_TYPE_VERTEX_BUFFER_NV :: IndirectCommandsTokenTypeNV

pattern INDIRECT_COMMANDS_TOKEN_TYPE_PUSH_CONSTANT_NV :: IndirectCommandsTokenTypeNV

pattern INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_INDEXED_NV :: IndirectCommandsTokenTypeNV

pattern INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_NV :: IndirectCommandsTokenTypeNV

pattern INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_TASKS_NV :: IndirectCommandsTokenTypeNV

type NV_DEVICE_GENERATED_COMMANDS_SPEC_VERSION = 3

pattern NV_DEVICE_GENERATED_COMMANDS_SPEC_VERSION :: forall a. Integral a => a

type NV_DEVICE_GENERATED_COMMANDS_EXTENSION_NAME = "VK_NV_device_generated_commands"

pattern NV_DEVICE_GENERATED_COMMANDS_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype IndirectCommandsLayoutNV

VkIndirectCommandsLayoutNV - Opaque handle to an indirect commands layout object

See Also

GeneratedCommandsInfoNV, GeneratedCommandsMemoryRequirementsInfoNV, createIndirectCommandsLayoutNV, destroyIndirectCommandsLayoutNV

Constructors

IndirectCommandsLayoutNV Word64