Vulkan.Extensions.VK_NV_external_memory

Description

Name

VK_NV_external_memory - device extension

VK_NV_external_memory

Name String

VK_NV_external_memory

Extension Type

Device extension

Registered Extension Number

57

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_NV_external_memory_capabilities

Deprecation state

Deprecated by VK_KHR_external_memory extension

Which in turn was promoted to Vulkan 1.1

Contact

James Jones https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_NV_external_memory:%20&body=@cubanismo%20

Other Extension Metadata

Last Modified Date

2016-08-19

IP Status

No known IP claims.

Contributors

James Jones, NVIDIA

Carsten Rohde, NVIDIA

Description

Applications may wish to export memory to other Vulkan instances or other APIs, or import memory from other Vulkan instances or other APIs to enable Vulkan workloads to be split up across application module, process, or API boundaries. This extension enables applications to create exportable Vulkan memory objects such that the underlying resources can be referenced outside the Vulkan instance that created them.

New Structures

Extending ImageCreateInfo:

ExternalMemoryImageCreateInfoNV

Extending MemoryAllocateInfo:

ExportMemoryAllocateInfoNV

New Enum Constants

NV_EXTERNAL_MEMORY_EXTENSION_NAME

NV_EXTERNAL_MEMORY_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_EXPORT_MEMORY_ALLOCATE_INFO_NV

STRUCTURE_TYPE_EXTERNAL_MEMORY_IMAGE_CREATE_INFO_NV

Issues

1) If memory objects are shared between processes and APIs, is this considered aliasing according to the rules outlined in the Memory Aliasing section?

RESOLVED: Yes, but strict exceptions to the rules are added to allow some forms of aliasing in these cases. Further, other extensions may build upon these new aliasing rules to define specific support usage within Vulkan for imported native memory objects, or memory objects from other APIs.

2) Are new image layouts or metadata required to specify image layouts and layout transitions compatible with non-Vulkan APIs, or with other instances of the same Vulkan driver?

RESOLVED: No. Separate instances of the same Vulkan driver running on the same GPU should have identical internal layout semantics, so applictions have the tools they need to ensure views of images are consistent between the two instances. Other APIs will fall into two categories: Those that are Vulkan compatible (a term to be defined by subsequent interopability extensions), or Vulkan incompatible. When sharing images with Vulkan incompatible APIs, the Vulkan image must be transitioned to the IMAGE_LAYOUT_GENERAL layout before handing it off to the external API.

Note this does not attempt to address cross-device transitions, nor transitions to engines on the same device which are not visible within the Vulkan API. Both of these are beyond the scope of this extension.

Examples

    // TODO: Write some sample code here.

Version History

Revision 1, 2016-08-19 (James Jones)

Initial draft

See Also

ExportMemoryAllocateInfoNV, ExternalMemoryImageCreateInfoNV

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data ExternalMemoryImageCreateInfoNV

VkExternalMemoryImageCreateInfoNV - Specify that an image may be backed by external memory

Valid Usage (Implicit)

See Also

ExternalMemoryHandleTypeFlagsNV, StructureType

Constructors

ExternalMemoryImageCreateInfoNV

Fields

handleTypes :: ExternalMemoryHandleTypeFlagsNV

handleTypes is zero, or a bitmask of ExternalMemoryHandleTypeFlagBitsNV specifying one or more external memory handle types.

handleTypes must be a valid combination of ExternalMemoryHandleTypeFlagBitsNV values

data ExportMemoryAllocateInfoNV

VkExportMemoryAllocateInfoNV - Specify memory handle types that may be exported

Valid Usage (Implicit)

See Also

ExternalMemoryHandleTypeFlagsNV, StructureType

Constructors

ExportMemoryAllocateInfoNV

Fields

handleTypes :: ExternalMemoryHandleTypeFlagsNV

handleTypes is a bitmask of ExternalMemoryHandleTypeFlagBitsNV specifying one or more memory handle types that may be exported. Multiple handle types may be requested for the same allocation as long as they are compatible, as reported by getPhysicalDeviceExternalImageFormatPropertiesNV.

handleTypes must be a valid combination of ExternalMemoryHandleTypeFlagBitsNV values

type NV_EXTERNAL_MEMORY_SPEC_VERSION = 1

pattern NV_EXTERNAL_MEMORY_SPEC_VERSION :: forall a. Integral a => a

type NV_EXTERNAL_MEMORY_EXTENSION_NAME = "VK_NV_external_memory"

pattern NV_EXTERNAL_MEMORY_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype ExternalMemoryHandleTypeFlagBitsNV

VkExternalMemoryHandleTypeFlagBitsNV - Bitmask specifying external memory handle types

See Also

ExternalMemoryHandleTypeFlagsNV

Constructors

ExternalMemoryHandleTypeFlagBitsNV Flags

Bundled Patterns

pattern EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT_NV specifies a handle to memory returned by getMemoryWin32HandleNV, or one duplicated from such a handle using DuplicateHandle().

pattern EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_NV specifies a handle to memory returned by getMemoryWin32HandleNV.

pattern EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_BIT_NV specifies a valid NT handle to memory returned by IDXGIResource1::CreateSharedHandle, or a handle duplicated from such a handle using DuplicateHandle().

pattern EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_KMT_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_KMT_BIT_NV specifies a handle to memory returned by IDXGIResource::GetSharedHandle().

type ExternalMemoryHandleTypeFlagsNV = ExternalMemoryHandleTypeFlagBitsNV