Vulkan.Extensions.VK_NV_external_memory_capabilities

Description

Name

VK_NV_external_memory_capabilities - instance extension

VK_NV_external_memory_capabilities

Name String

VK_NV_external_memory_capabilities

Extension Type

Instance extension

Registered Extension Number

56

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Deprecation state

Deprecated by VK_KHR_external_memory_capabilities extension

Which in turn was promoted to Vulkan 1.1

Contact

James Jones https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_NV_external_memory_capabilities:%20&body=@cubanismo%20

Other Extension Metadata

Last Modified Date

2016-08-19

IP Status

No known IP claims.

Interactions and External Dependencies

Interacts with Vulkan 1.1.

Interacts with VK_KHR_dedicated_allocation.

Interacts with VK_NV_dedicated_allocation.

Contributors

James Jones, NVIDIA

Description

Applications may wish to import memory from the Direct 3D API, or export memory to other Vulkan instances. This extension provides a set of capability queries that allow applications determine what types of win32 memory handles an implementation supports for a given set of use cases.

New Commands

getPhysicalDeviceExternalImageFormatPropertiesNV

New Structures

ExternalImageFormatPropertiesNV

New Enums

ExternalMemoryFeatureFlagBitsNV

ExternalMemoryHandleTypeFlagBitsNV

New Bitmasks

ExternalMemoryFeatureFlagsNV

ExternalMemoryHandleTypeFlagsNV

New Enum Constants

NV_EXTERNAL_MEMORY_CAPABILITIES_EXTENSION_NAME

NV_EXTERNAL_MEMORY_CAPABILITIES_SPEC_VERSION

Issues

1) Why do so many external memory capabilities need to be queried on a per-memory-handle-type basis?

RESOLVED: This is because some handle types are based on OS-native objects that have far more limited capabilities than the very generic Vulkan memory objects. Not all memory handle types can name memory objects that support 3D images, for example. Some handle types cannot even support the deferred image and memory binding behavior of Vulkan and require specifying the image when allocating or importing the memory object.

2) Does the ExternalImageFormatPropertiesNV struct need to include a list of memory type bits that support the given handle type?

RESOLVED: No. The memory types that do not support the handle types will simply be filtered out of the results returned by getImageMemoryRequirements when a set of handle types was specified at image creation time.

3) Should the non-opaque handle types be moved to their own extension?

RESOLVED: Perhaps. However, defining the handle type bits does very little and does not require any platform-specific types on its own, and it is easier to maintain the bitmask values in a single extension for now. Presumably more handle types could be added by separate extensions though, and it would be midly weird to have some platform-specific ones defined in the core spec and some in extensions

Version History

Revision 1, 2016-08-19 (James Jones)

Initial version

See Also

ExternalImageFormatPropertiesNV, ExternalMemoryFeatureFlagBitsNV, ExternalMemoryFeatureFlagsNV, ExternalMemoryHandleTypeFlagBitsNV, ExternalMemoryHandleTypeFlagsNV, getPhysicalDeviceExternalImageFormatPropertiesNV

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

getPhysicalDeviceExternalImageFormatPropertiesNV

Arguments

:: forall io. MonadIO io

=> PhysicalDevice
physicalDevice is the physical device from which to query the image capabilities

physicalDevice must be a valid PhysicalDevice handle

-> Format
format is the image format, corresponding to ImageCreateInfo::format.

format must be a valid Format value

-> ImageType
type is the image type, corresponding to ImageCreateInfo::imageType.

type must be a valid ImageType value

-> ImageTiling
tiling is the image tiling, corresponding to ImageCreateInfo::tiling.

tiling must be a valid ImageTiling value

-> ImageUsageFlags
usage is the intended usage of the image, corresponding to ImageCreateInfo::usage.

usage must be a valid combination of ImageUsageFlagBits values

usage must not be 0

-> ImageCreateFlags
flags is a bitmask describing additional parameters of the image, corresponding to ImageCreateInfo::flags.

flags must be a valid combination of ImageCreateFlagBits values

-> ("externalHandleType" ::: ExternalMemoryHandleTypeFlagsNV)
externalHandleType is either one of the bits from ExternalMemoryHandleTypeFlagBitsNV, or 0.

externalHandleType must be a valid combination of ExternalMemoryHandleTypeFlagBitsNV values

-> io ExternalImageFormatPropertiesNV

vkGetPhysicalDeviceExternalImageFormatPropertiesNV - determine image capabilities compatible with external memory handle types

Description

If externalHandleType is 0, pExternalImageFormatProperties->imageFormatProperties will return the same values as a call to getPhysicalDeviceImageFormatProperties, and the other members of pExternalImageFormatProperties will all be 0. Otherwise, they are filled in as described for ExternalImageFormatPropertiesNV.

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_FORMAT_NOT_SUPPORTED

See Also

ExternalImageFormatPropertiesNV, ExternalMemoryHandleTypeFlagsNV, Format, ImageCreateFlags, ImageTiling, ImageType, ImageUsageFlags, PhysicalDevice

data ExternalImageFormatPropertiesNV

VkExternalImageFormatPropertiesNV - Structure specifying external image format properties

See Also

ExternalMemoryFeatureFlagsNV, ExternalMemoryHandleTypeFlagsNV, ImageFormatProperties, getPhysicalDeviceExternalImageFormatPropertiesNV

Constructors

ExternalImageFormatPropertiesNV

Fields

imageFormatProperties :: ImageFormatProperties

imageFormatProperties will be filled in as when calling getPhysicalDeviceImageFormatProperties, but the values returned may vary depending on the external handle type requested.

externalMemoryFeatures :: ExternalMemoryFeatureFlagsNV

externalMemoryFeatures is a bitmask of ExternalMemoryFeatureFlagBitsNV, indicating properties of the external memory handle type (getPhysicalDeviceExternalImageFormatPropertiesNV::externalHandleType) being queried, or 0 if the external memory handle type is 0.

exportFromImportedHandleTypes :: ExternalMemoryHandleTypeFlagsNV

exportFromImportedHandleTypes is a bitmask of ExternalMemoryHandleTypeFlagBitsNV containing a bit set for every external handle type that may be used to create memory from which the handles of the type specified in getPhysicalDeviceExternalImageFormatPropertiesNV::externalHandleType can be exported, or 0 if the external memory handle type is 0.

compatibleHandleTypes :: ExternalMemoryHandleTypeFlagsNV

compatibleHandleTypes is a bitmask of ExternalMemoryHandleTypeFlagBitsNV containing a bit set for every external handle type that may be specified simultaneously with the handle type specified by getPhysicalDeviceExternalImageFormatPropertiesNV::externalHandleType when calling allocateMemory, or 0 if the external memory handle type is 0. compatibleHandleTypes will always contain getPhysicalDeviceExternalImageFormatPropertiesNV::externalHandleType

type ExternalMemoryHandleTypeFlagsNV = ExternalMemoryHandleTypeFlagBitsNV

newtype ExternalMemoryHandleTypeFlagBitsNV

VkExternalMemoryHandleTypeFlagBitsNV - Bitmask specifying external memory handle types

See Also

ExternalMemoryHandleTypeFlagsNV

Constructors

ExternalMemoryHandleTypeFlagBitsNV Flags

Bundled Patterns

pattern EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT_NV specifies a handle to memory returned by getMemoryWin32HandleNV, or one duplicated from such a handle using DuplicateHandle().

pattern EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_NV specifies a handle to memory returned by getMemoryWin32HandleNV.

pattern EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_BIT_NV specifies a valid NT handle to memory returned by IDXGIResource1::CreateSharedHandle, or a handle duplicated from such a handle using DuplicateHandle().

pattern EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_KMT_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_KMT_BIT_NV specifies a handle to memory returned by IDXGIResource::GetSharedHandle().

type ExternalMemoryFeatureFlagsNV = ExternalMemoryFeatureFlagBitsNV

newtype ExternalMemoryFeatureFlagBitsNV

VkExternalMemoryFeatureFlagBitsNV - Bitmask specifying external memory features

See Also

ExternalImageFormatPropertiesNV, ExternalMemoryFeatureFlagsNV, getPhysicalDeviceExternalImageFormatPropertiesNV

Constructors

ExternalMemoryFeatureFlagBitsNV Flags

Bundled Patterns

pattern EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT_NV :: ExternalMemoryFeatureFlagBitsNV
EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT_NV specifies that external memory of the specified type must be created as a dedicated allocation when used in the manner specified.

pattern EXTERNAL_MEMORY_FEATURE_EXPORTABLE_BIT_NV :: ExternalMemoryFeatureFlagBitsNV
EXTERNAL_MEMORY_FEATURE_EXPORTABLE_BIT_NV specifies that the implementation supports exporting handles of the specified type.

pattern EXTERNAL_MEMORY_FEATURE_IMPORTABLE_BIT_NV :: ExternalMemoryFeatureFlagBitsNV
EXTERNAL_MEMORY_FEATURE_IMPORTABLE_BIT_NV specifies that the implementation supports importing handles of the specified type.

type NV_EXTERNAL_MEMORY_CAPABILITIES_SPEC_VERSION = 1

pattern NV_EXTERNAL_MEMORY_CAPABILITIES_SPEC_VERSION :: forall a. Integral a => a

type NV_EXTERNAL_MEMORY_CAPABILITIES_EXTENSION_NAME = "VK_NV_external_memory_capabilities"

pattern NV_EXTERNAL_MEMORY_CAPABILITIES_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a