Vulkan.Extensions.VK_NV_fill_rectangle

Description

Name

VK_NV_fill_rectangle - device extension

VK_NV_fill_rectangle

Name String

VK_NV_fill_rectangle

Extension Type

Device extension

Registered Extension Number

154

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Contact

Jeff Bolz https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_NV_fill_rectangle:%20&body=@jeffbolznv%20

Other Extension Metadata

Last Modified Date

2017-05-22

Contributors

Jeff Bolz, NVIDIA

Description

This extension adds a new PolygonMode enum where a triangle is rasterized by computing and filling its axis-aligned screen-space bounding box, disregarding the actual triangle edges. This can be useful for drawing a rectangle without being split into two triangles with an internal edge. It is also useful to minimize the number of primitives that need to be drawn, particularly for a user interface.

New Enum Constants

NV_FILL_RECTANGLE_EXTENSION_NAME

NV_FILL_RECTANGLE_SPEC_VERSION

Extending PolygonMode:

POLYGON_MODE_FILL_RECTANGLE_NV

Version History

Revision 1, 2017-05-22 (Jeff Bolz)

Internal revisions

See Also

No cross-references are available

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

type NV_FILL_RECTANGLE_SPEC_VERSION = 1

pattern NV_FILL_RECTANGLE_SPEC_VERSION :: forall a. Integral a => a

type NV_FILL_RECTANGLE_EXTENSION_NAME = "VK_NV_fill_rectangle"

pattern NV_FILL_RECTANGLE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a