Vulkan.Extensions.VK_NV_fragment_coverage_to_color

Description

Name

VK_NV_fragment_coverage_to_color - device extension

VK_NV_fragment_coverage_to_color

Name String

VK_NV_fragment_coverage_to_color

Extension Type

Device extension

Registered Extension Number

150

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Contact

Jeff Bolz https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_NV_fragment_coverage_to_color:%20&body=@jeffbolznv%20

Other Extension Metadata

Last Modified Date

2017-05-21

Contributors

Jeff Bolz, NVIDIA

Description

This extension allows the fragment coverage value, represented as an integer bitmask, to be substituted for a color output being written to a single-component color attachment with integer components (e.g. FORMAT_R8_UINT). The functionality provided by this extension is different from simply writing the SampleMask fragment shader output, in that the coverage value written to the framebuffer is taken after stencil test and depth test, as well as after fragment operations such as alpha-to-coverage.

This functionality may be useful for deferred rendering algorithms, where the second pass needs to know which samples belong to which original fragments.

New Structures

Extending PipelineMultisampleStateCreateInfo:

PipelineCoverageToColorStateCreateInfoNV

New Bitmasks

PipelineCoverageToColorStateCreateFlagsNV

New Enum Constants

NV_FRAGMENT_COVERAGE_TO_COLOR_EXTENSION_NAME

NV_FRAGMENT_COVERAGE_TO_COLOR_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_PIPELINE_COVERAGE_TO_COLOR_STATE_CREATE_INFO_NV

Version History

Revision 1, 2017-05-21 (Jeff Bolz)

Internal revisions

See Also

PipelineCoverageToColorStateCreateFlagsNV, PipelineCoverageToColorStateCreateInfoNV

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data PipelineCoverageToColorStateCreateInfoNV

VkPipelineCoverageToColorStateCreateInfoNV - Structure specifying whether fragment coverage replaces a color

Description

If the pNext chain of PipelineMultisampleStateCreateInfo includes a PipelineCoverageToColorStateCreateInfoNV structure, then that structure controls whether the fragment coverage is substituted for a fragment color output and, if so, which output is replaced.

If coverageToColorEnable is TRUE, the coverage mask replaces the first component of the color value corresponding to the fragment shader output location with Location equal to coverageToColorLocation and Index equal to zero. If the color attachment format has fewer bits than the coverage mask, the low bits of the sample coverage mask are taken without any clamping. If the color attachment format has more bits than the coverage mask, the high bits of the sample coverage mask are filled with zeros.

If coverageToColorEnable is FALSE, these operations are skipped. If this structure is not present, it is as if coverageToColorEnable is FALSE.

Valid Usage

If coverageToColorEnable is TRUE, then the render pass subpass indicated by GraphicsPipelineCreateInfo::renderPass and GraphicsPipelineCreateInfo::subpass must have a color attachment at the location selected by coverageToColorLocation, with a Format of FORMAT_R8_UINT, FORMAT_R8_SINT, FORMAT_R16_UINT, FORMAT_R16_SINT, FORMAT_R32_UINT, or FORMAT_R32_SINT

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PIPELINE_COVERAGE_TO_COLOR_STATE_CREATE_INFO_NV

flags must be 0

See Also

Bool32, PipelineCoverageToColorStateCreateFlagsNV, StructureType

Constructors

PipelineCoverageToColorStateCreateInfoNV

Fields

flags :: PipelineCoverageToColorStateCreateFlagsNV

flags is reserved for future use.

coverageToColorEnable :: Bool

coverageToColorEnable controls whether the fragment coverage value replaces a fragment color output.

coverageToColorLocation :: Word32

coverageToColorLocation controls which fragment shader color output value is replaced.

newtype PipelineCoverageToColorStateCreateFlagsNV

VkPipelineCoverageToColorStateCreateFlagsNV - Reserved for future use

Description

PipelineCoverageToColorStateCreateFlagsNV is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

PipelineCoverageToColorStateCreateInfoNV

Constructors

PipelineCoverageToColorStateCreateFlagsNV Flags

type NV_FRAGMENT_COVERAGE_TO_COLOR_SPEC_VERSION = 1

pattern NV_FRAGMENT_COVERAGE_TO_COLOR_SPEC_VERSION :: forall a. Integral a => a

type NV_FRAGMENT_COVERAGE_TO_COLOR_EXTENSION_NAME = "VK_NV_fragment_coverage_to_color"

pattern NV_FRAGMENT_COVERAGE_TO_COLOR_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a