Vulkan.Extensions.VK_NV_representative_fragment_test

Documentation

data PhysicalDeviceRepresentativeFragmentTestFeaturesNV

VkPhysicalDeviceRepresentativeFragmentTestFeaturesNV - Structure describing the representative fragment test features that can be supported by an implementation

Members

The members of the PhysicalDeviceRepresentativeFragmentTestFeaturesNV structure describe the following features:

Description

If the PhysicalDeviceRepresentativeFragmentTestFeaturesNV structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether the feature is supported. PhysicalDeviceRepresentativeFragmentTestFeaturesNV can also be included in the pNext chain of DeviceCreateInfo to enable the feature.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceRepresentativeFragmentTestFeaturesNV

Fields

representativeFragmentTest :: Bool

representativeFragmentTest indicates whether the implementation supports the representative fragment test. See Representative Fragment Test.

data PipelineRepresentativeFragmentTestStateCreateInfoNV

VkPipelineRepresentativeFragmentTestStateCreateInfoNV - Structure specifying representative fragment test

Description

If this structure is not present, representativeFragmentTestEnable is considered to be FALSE, and the representative fragment test is disabled.

If early fragment tests are not enabled in the active fragment shader, the representative fragment shader test has no effect, even if enabled.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PipelineRepresentativeFragmentTestStateCreateInfoNV

Fields

representativeFragmentTestEnable :: Bool

representativeFragmentTestEnable controls whether the representative fragment test is enabled.

type NV_REPRESENTATIVE_FRAGMENT_TEST_SPEC_VERSION = 2

pattern NV_REPRESENTATIVE_FRAGMENT_TEST_SPEC_VERSION :: forall a. Integral a => a

type NV_REPRESENTATIVE_FRAGMENT_TEST_EXTENSION_NAME = "VK_NV_representative_fragment_test"

pattern NV_REPRESENTATIVE_FRAGMENT_TEST_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a