Vulkan.Extensions.VK_NV_shader_subgroup_partitioned

Description

Name

VK_NV_shader_subgroup_partitioned - device extension

VK_NV_shader_subgroup_partitioned

Name String

VK_NV_shader_subgroup_partitioned

Extension Type

Device extension

Registered Extension Number

199

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.1

Contact

Jeff Bolz https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_NV_shader_subgroup_partitioned:%20&body=@jeffbolznv%20

Other Extension Metadata

Last Modified Date

2018-03-17

Interactions and External Dependencies

This extension requires SPV_NV_shader_subgroup_partitioned

This extension provides API support for GL_NV_shader_subgroup_partitioned

Contributors

Jeff Bolz, NVIDIA

Description

This extension enables support for a new class of group operations on subgroups via the GL_NV_shader_subgroup_partitioned GLSL extension and SPV_NV_shader_subgroup_partitioned SPIR-V extension. Support for these new operations is advertised via the SUBGROUP_FEATURE_PARTITIONED_BIT_NV bit.

This extension requires Vulkan 1.1, for general subgroup support.

New Enum Constants

NV_SHADER_SUBGROUP_PARTITIONED_EXTENSION_NAME

NV_SHADER_SUBGROUP_PARTITIONED_SPEC_VERSION

Extending SubgroupFeatureFlagBits:

SUBGROUP_FEATURE_PARTITIONED_BIT_NV

Version History

Revision 1, 2018-03-17 (Jeff Bolz)

Internal revisions

See Also

No cross-references are available

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

type NV_SHADER_SUBGROUP_PARTITIONED_SPEC_VERSION = 1

pattern NV_SHADER_SUBGROUP_PARTITIONED_SPEC_VERSION :: forall a. Integral a => a

type NV_SHADER_SUBGROUP_PARTITIONED_EXTENSION_NAME = "VK_NV_shader_subgroup_partitioned"

pattern NV_SHADER_SUBGROUP_PARTITIONED_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a