Vulkan.Extensions.VK_NV_shading_rate_image

Description

Name

VK_NV_shading_rate_image - device extension

VK_NV_shading_rate_image

Name String

VK_NV_shading_rate_image

Extension Type

Device extension

Registered Extension Number

165

Revision

3

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_get_physical_device_properties2

Contact

Pat Brown https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_NV_shading_rate_image:%20&body=@nvpbrown%20

Other Extension Metadata

Last Modified Date

2019-07-18

Interactions and External Dependencies

This extension requires SPV_NV_shading_rate

This extension provides API support for GL_NV_shading_rate_image

Contributors

Pat Brown, NVIDIA

Carsten Rohde, NVIDIA

Jeff Bolz, NVIDIA

Daniel Koch, NVIDIA

Mathias Schott, NVIDIA

Matthew Netsch, Qualcomm Technologies, Inc.

Description

This extension allows applications to use a variable shading rate when processing fragments of rasterized primitives. By default, Vulkan will spawn one fragment shader for each pixel covered by a primitive. In this extension, applications can bind a shading rate image that can be used to vary the number of fragment shader invocations across the framebuffer. Some portions of the screen may be configured to spawn up to 16 fragment shaders for each pixel, while other portions may use a single fragment shader invocation for a 4x4 block of pixels. This can be useful for use cases like eye tracking, where the portion of the framebuffer that the user is looking at directly can be processed at high frequency, while distant corners of the image can be processed at lower frequency. Each texel in the shading rate image represents a fixed-size rectangle in the framebuffer, covering 16x16 pixels in the initial implementation of this extension. When rasterizing a primitive covering one of these rectangles, the Vulkan implementation reads a texel in the bound shading rate image and looks up the fetched value in a palette to determine a base shading rate.

In addition to the API support controlling rasterization, this extension also adds Vulkan support for the SPV_NV_shading_rate extension to SPIR-V. That extension provides two fragment shader variable decorations that allow fragment shaders to determine the shading rate used for processing the fragment:

FragmentSizeNV, which indicates the width and height of the set of pixels processed by the fragment shader.

InvocationsPerPixel, which indicates the maximum number of fragment shader invocations that could be spawned for the pixel(s) covered by the fragment.

When using SPIR-V in conjunction with the OpenGL Shading Language (GLSL), the fragment shader capabilities are provided by the GL_NV_shading_rate_image language extension and correspond to the built-in variables gl_FragmentSizeNV and gl_InvocationsPerPixelNV, respectively.

New Commands

cmdBindShadingRateImageNV

cmdSetCoarseSampleOrderNV

cmdSetViewportShadingRatePaletteNV

New Structures

CoarseSampleLocationNV

CoarseSampleOrderCustomNV

ShadingRatePaletteNV

Extending PhysicalDeviceFeatures2, DeviceCreateInfo:

PhysicalDeviceShadingRateImageFeaturesNV

Extending PhysicalDeviceProperties2:

PhysicalDeviceShadingRateImagePropertiesNV

Extending PipelineViewportStateCreateInfo:

PipelineViewportCoarseSampleOrderStateCreateInfoNV

PipelineViewportShadingRateImageStateCreateInfoNV

New Enums

CoarseSampleOrderTypeNV

ShadingRatePaletteEntryNV

New Enum Constants

NV_SHADING_RATE_IMAGE_EXTENSION_NAME

NV_SHADING_RATE_IMAGE_SPEC_VERSION

Extending AccessFlagBits:

ACCESS_SHADING_RATE_IMAGE_READ_BIT_NV

Extending DynamicState:

DYNAMIC_STATE_VIEWPORT_COARSE_SAMPLE_ORDER_NV

DYNAMIC_STATE_VIEWPORT_SHADING_RATE_PALETTE_NV

Extending ImageLayout:

IMAGE_LAYOUT_SHADING_RATE_OPTIMAL_NV

Extending ImageUsageFlagBits:

IMAGE_USAGE_SHADING_RATE_IMAGE_BIT_NV

Extending PipelineStageFlagBits:

PIPELINE_STAGE_SHADING_RATE_IMAGE_BIT_NV

Extending StructureType:

STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADING_RATE_IMAGE_FEATURES_NV

STRUCTURE_TYPE_PHYSICAL_DEVICE_SHADING_RATE_IMAGE_PROPERTIES_NV

STRUCTURE_TYPE_PIPELINE_VIEWPORT_COARSE_SAMPLE_ORDER_STATE_CREATE_INFO_NV

STRUCTURE_TYPE_PIPELINE_VIEWPORT_SHADING_RATE_IMAGE_STATE_CREATE_INFO_NV

Issues

When using shading rates specifying “coarse” fragments covering multiple pixels, we will generate a combined coverage mask that combines the coverage masks of all pixels covered by the fragment. By default, these masks are combined in an implementation-dependent order. Should we provide a mechanism allowing applications to query or specify an exact order?

RESOLVED: Yes, this feature is useful for cases where most of the fragment shader can be evaluated once for an entire coarse fragment, but where some per-pixel computations are also required. For example, a per-pixel alpha test may want to kill all the samples for some pixels in a coarse fragment. This sort of test can be implemented using an output sample mask, but such a shader would need to know which bit in the mask corresponds to each sample in the coarse fragment. We are including a mechanism to allow aplications to specify the orders of coverage samples for each shading rate and sample count, either as static pipeline state or dynamically via a command buffer. This portion of the extension has its own feature bit.

We will not be providing a query to determine the implementation-dependent default ordering. The thinking here is that if an application cares enough about the coarse fragment sample ordering to perform such a query, it could instead just set its own order, also using custom per-pixel sample locations if required.

For the pipeline stage PIPELINE_STAGE_SHADING_RATE_IMAGE_BIT_NV, should we specify a precise location in the pipeline the shading rate image is accessed (after geometry shading, but before the early fragment tests) or leave it under-specified in case there are other implementations that access the image in a different pipeline location?

RESOLVED We are specifying the pipeline stage to be between the final stage used for vertex processing (PIPELINE_STAGE_GEOMETRY_SHADER_BIT) and before the first stage used for fragment processing (PIPELINE_STAGE_EARLY_FRAGMENT_TESTS_BIT), which seems to be the natural place to access the shading rate image.

How do centroid-sampled variables work with fragments larger than one pixel?

RESOLVED For single-pixel fragments, fragment shader inputs decorated with Centroid are sampled at an implementation-dependent location in the intersection of the area of the primitive being rasterized and the area of the pixel that corresponds to the fragment. With multi-pixel fragments, we follow a similar pattern, using the intersection of the primitive and the set of pixels corresponding to the fragment.

One important thing to keep in mind when using such “coarse” shading rates is that fragment attributes are sampled at the center of the fragment by default, regardless of the set of pixels/samples covered by the fragment. For fragments with a size of 4x4 pixels, this center location will be more than two pixels (1.5 * sqrt(2)) away from the center of the pixels at the corners of the fragment. When rendering a primitive that covers only a small part of a coarse fragment, sampling a color outside the primitive can produce overly bright or dark color values if the color values have a large gradient. To deal with this, an application can use centroid sampling on attributes where “extrapolation” artifacts can lead to overly bright or dark pixels. Note that this same problem also exists for multisampling with single-pixel fragments, but is less severe because it only affects certain samples of a pixel and such bright/dark samples may be averaged with other samples that don’t have a similar problem.

Version History

Revision 3, 2019-07-18 (Mathias Schott)

Fully list extension interfaces in this appendix.

Revision 2, 2018-09-13 (Pat Brown)

Miscellaneous edits preparing the specification for publication.

Revision 1, 2018-08-08 (Pat Brown)

Internal revisions

See Also

CoarseSampleLocationNV, CoarseSampleOrderCustomNV, CoarseSampleOrderTypeNV, PhysicalDeviceShadingRateImageFeaturesNV, PhysicalDeviceShadingRateImagePropertiesNV, PipelineViewportCoarseSampleOrderStateCreateInfoNV, PipelineViewportShadingRateImageStateCreateInfoNV, ShadingRatePaletteEntryNV, ShadingRatePaletteNV, cmdBindShadingRateImageNV, cmdSetCoarseSampleOrderNV, cmdSetViewportShadingRatePaletteNV

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

cmdBindShadingRateImageNV

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command will be recorded.

-> ImageView
imageView is an image view handle specifying the shading rate image. imageView may be set to NULL_HANDLE, which is equivalent to specifying a view of an image filled with zero values.

-> ImageLayout
imageLayout is the layout that the image subresources accessible from imageView will be in when the shading rate image is accessed.

-> io ()

vkCmdBindShadingRateImageNV - Bind a shading rate image on a command buffer

Valid Usage

The shading rate image feature must be enabled

If imageView is not NULL_HANDLE, it must be a valid ImageView handle of type IMAGE_VIEW_TYPE_2D or IMAGE_VIEW_TYPE_2D_ARRAY

If imageView is not NULL_HANDLE, it must have a format of FORMAT_R8_UINT

If imageView is not NULL_HANDLE, it must have been created with a usage value including IMAGE_USAGE_SHADING_RATE_IMAGE_BIT_NV

If imageView is not NULL_HANDLE, imageLayout must match the actual ImageLayout of each subresource accessible from imageView at the time the subresource is accessed

If imageView is not NULL_HANDLE, imageLayout must be IMAGE_LAYOUT_SHADING_RATE_OPTIMAL_NV or IMAGE_LAYOUT_GENERAL

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

If imageView is not NULL_HANDLE, imageView must be a valid ImageView handle

imageLayout must be a valid ImageLayout value

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

Both of commandBuffer, and imageView that are valid handles of non-ignored parameters must have been created, allocated, or retrieved from the same Device

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

CommandBuffer, ImageLayout, ImageView

cmdSetViewportShadingRatePaletteNV

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command will be recorded.

-> ("firstViewport" ::: Word32)
firstViewport is the index of the first viewport whose shading rate palette is updated by the command.

-> ("shadingRatePalettes" ::: Vector ShadingRatePaletteNV)
pShadingRatePalettes is a pointer to an array of ShadingRatePaletteNV structures defining the palette for each viewport.

-> io ()

vkCmdSetViewportShadingRatePaletteNV - Set shading rate image palettes on a command buffer

Valid Usage

The shading rate image feature must be enabled

The sum of firstViewport and viewportCount must be between 1 and PhysicalDeviceLimits::maxViewports, inclusive

If the multiple viewports feature is not enabled, firstViewport must be 0

If the multiple viewports feature is not enabled, viewportCount must be 1

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pShadingRatePalettes must be a valid pointer to an array of viewportCount valid ShadingRatePaletteNV structures

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

viewportCount must be greater than 0

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

CommandBuffer, ShadingRatePaletteNV

cmdSetCoarseSampleOrderNV

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command will be recorded.

-> CoarseSampleOrderTypeNV
sampleOrderType specifies the mechanism used to order coverage samples in fragments larger than one pixel.

-> ("customSampleOrders" ::: Vector CoarseSampleOrderCustomNV)
pCustomSampleOrders is a pointer to an array of CoarseSampleOrderCustomNV structures, each of which specifies the coverage sample order for a single combination of fragment area and coverage sample count.

-> io ()

vkCmdSetCoarseSampleOrderNV - Set sample order for coarse fragments on a command buffer

Description

If sampleOrderType is COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV, the coverage sample order used for any combination of fragment area and coverage sample count not enumerated in pCustomSampleOrders will be identical to that used for COARSE_SAMPLE_ORDER_TYPE_DEFAULT_NV.

Valid Usage

If sampleOrderType is not COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV, customSamplerOrderCount must be 0

The array pCustomSampleOrders must not contain two structures with matching values for both the shadingRate and sampleCount members

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

sampleOrderType must be a valid CoarseSampleOrderTypeNV value

If customSampleOrderCount is not 0, pCustomSampleOrders must be a valid pointer to an array of customSampleOrderCount valid CoarseSampleOrderCustomNV structures

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

CoarseSampleOrderCustomNV, CoarseSampleOrderTypeNV, CommandBuffer

data ShadingRatePaletteNV

VkShadingRatePaletteNV - Structure specifying a single shading rate palette

Valid Usage (Implicit)

See Also

PipelineViewportShadingRateImageStateCreateInfoNV, ShadingRatePaletteEntryNV, cmdSetViewportShadingRatePaletteNV

Constructors

ShadingRatePaletteNV

Fields

shadingRatePaletteEntries :: Vector ShadingRatePaletteEntryNV

pShadingRatePaletteEntries is a pointer to an array of ShadingRatePaletteEntryNV enums defining the shading rate for each palette entry.

pShadingRatePaletteEntries must be a valid pointer to an array of shadingRatePaletteEntryCount valid ShadingRatePaletteEntryNV values

data PipelineViewportShadingRateImageStateCreateInfoNV

VkPipelineViewportShadingRateImageStateCreateInfoNV - Structure specifying parameters controlling shading rate image usage

Description

If this structure is not present, shadingRateImageEnable is considered to be FALSE, and the shading rate image and palettes are not used.

Valid Usage

If the multiple viewports feature is not enabled, viewportCount must be 0 or 1

viewportCount must be less than or equal to PhysicalDeviceLimits::maxViewports

If shadingRateImageEnable is TRUE, viewportCount must be greater or equal to the viewportCount member of PipelineViewportStateCreateInfo

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PIPELINE_VIEWPORT_SHADING_RATE_IMAGE_STATE_CREATE_INFO_NV

viewportCount must be greater than 0

See Also

Bool32, ShadingRatePaletteNV, StructureType

Constructors

PipelineViewportShadingRateImageStateCreateInfoNV

Fields

shadingRateImageEnable :: Bool

shadingRateImageEnable specifies whether shading rate image and palettes are used during rasterization.

shadingRatePalettes :: Vector ShadingRatePaletteNV

pShadingRatePalettes is a pointer to an array of ShadingRatePaletteNV structures defining the palette for each viewport. If the shading rate palette state is dynamic, this member is ignored.

data PhysicalDeviceShadingRateImageFeaturesNV

VkPhysicalDeviceShadingRateImageFeaturesNV - Structure describing shading rate image features that can be supported by an implementation

Members

The members of the PhysicalDeviceShadingRateImageFeaturesNV structure describe the following features:

Description

See Shading Rate Image for more information.

If the PhysicalDeviceShadingRateImageFeaturesNV structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether the feature is supported. PhysicalDeviceShadingRateImageFeaturesNV can also be included in the pNext chain of DeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceShadingRateImageFeaturesNV

Fields

shadingRateImage :: Bool

shadingRateImage indicates that the implementation supports the use of a shading rate image to derive an effective shading rate for fragment processing. It also indicates that the implementation supports the ShadingRateNV SPIR-V execution mode.

shadingRateCoarseSampleOrder :: Bool

shadingRateCoarseSampleOrder indicates that the implementation supports a user-configurable ordering of coverage samples in fragments larger than one pixel.

data PhysicalDeviceShadingRateImagePropertiesNV

VkPhysicalDeviceShadingRateImagePropertiesNV - Structure describing shading rate image limits that can be supported by an implementation

Members

The members of the PhysicalDeviceShadingRateImagePropertiesNV structure describe the following implementation-dependent properties related to the shading rate image feature:

Description

If the PhysicalDeviceShadingRateImagePropertiesNV structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

Extent2D, StructureType

Constructors

PhysicalDeviceShadingRateImagePropertiesNV

Fields

shadingRateTexelSize :: Extent2D

shadingRateTexelSize indicates the width and height of the portion of the framebuffer corresponding to each texel in the shading rate image.

shadingRatePaletteSize :: Word32

shadingRatePaletteSize indicates the maximum number of palette entries supported for the shading rate image.

shadingRateMaxCoarseSamples :: Word32

shadingRateMaxCoarseSamples specifies the maximum number of coverage samples supported in a single fragment. If the product of the fragment size derived from the base shading rate and the number of coverage samples per pixel exceeds this limit, the final shading rate will be adjusted so that its product does not exceed the limit.

data CoarseSampleLocationNV

VkCoarseSampleLocationNV - Structure specifying parameters controlling shading rate image usage

Valid Usage

See Also

CoarseSampleOrderCustomNV

Constructors

CoarseSampleLocationNV

Fields

pixelX :: Word32

pixelX is added to the x coordinate of the upper-leftmost pixel of each fragment to identify the pixel containing the coverage sample.

pixelX must be less than the width (in pixels) of the fragment

pixelY :: Word32

pixelY is added to the y coordinate of the upper-leftmost pixel of each fragment to identify the pixel containing the coverage sample.

pixelY must be less than the height (in pixels) of the fragment

sample :: Word32

sample is the number of the coverage sample in the pixel identified by pixelX and pixelY.

sample must be less than the number of coverage samples in each pixel belonging to the fragment

data CoarseSampleOrderCustomNV

VkCoarseSampleOrderCustomNV - Structure specifying parameters controlling shading rate image usage

Description

When using a custom sample ordering, element j in pSampleLocations specifies a specific pixel location and sample index that corresponds to coverage index j in the multi-pixel fragment.

Valid Usage

shadingRate must be a shading rate that generates fragments with more than one pixel

sampleCount must correspond to a sample count enumerated in SampleCountFlags whose corresponding bit is set in PhysicalDeviceLimits::framebufferNoAttachmentsSampleCounts

sampleLocationCount must be equal to the product of sampleCount, the fragment width for shadingRate, and the fragment height for shadingRate

sampleLocationCount must be less than or equal to the value of PhysicalDeviceShadingRateImagePropertiesNV::shadingRateMaxCoarseSamples

The array pSampleLocations must contain exactly one entry for every combination of valid values for pixelX, pixelY, and sample in the structure CoarseSampleOrderCustomNV

Valid Usage (Implicit)

shadingRate must be a valid ShadingRatePaletteEntryNV value

pSampleLocations must be a valid pointer to an array of sampleLocationCount CoarseSampleLocationNV structures

sampleLocationCount must be greater than 0

See Also

CoarseSampleLocationNV, PipelineViewportCoarseSampleOrderStateCreateInfoNV, ShadingRatePaletteEntryNV, cmdSetCoarseSampleOrderNV

Constructors

CoarseSampleOrderCustomNV

Fields

shadingRate :: ShadingRatePaletteEntryNV

shadingRate is a shading rate palette entry that identifies the fragment width and height for the combination of fragment area and per-pixel coverage sample count to control.

sampleCount :: Word32

sampleCount identifies the per-pixel coverage sample count for the combination of fragment area and coverage sample count to control.

sampleLocations :: Vector CoarseSampleLocationNV

pSampleLocations is a pointer to an array of CoarseSampleOrderCustomNV structures specifying the location of each sample in the custom ordering.

data PipelineViewportCoarseSampleOrderStateCreateInfoNV

VkPipelineViewportCoarseSampleOrderStateCreateInfoNV - Structure specifying parameters controlling sample order in coarse fragments

Description

If this structure is not present, sampleOrderType is considered to be COARSE_SAMPLE_ORDER_TYPE_DEFAULT_NV.

If sampleOrderType is COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV, the coverage sample order used for any combination of fragment area and coverage sample count not enumerated in pCustomSampleOrders will be identical to that used for COARSE_SAMPLE_ORDER_TYPE_DEFAULT_NV.

If the pipeline was created with DYNAMIC_STATE_VIEWPORT_COARSE_SAMPLE_ORDER_NV, the contents of this structure (if present) are ignored, and the coverage sample order is instead specified by cmdSetCoarseSampleOrderNV.

Valid Usage

If sampleOrderType is not COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV, customSamplerOrderCount must be 0

The array pCustomSampleOrders must not contain two structures with matching values for both the shadingRate and sampleCount members

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PIPELINE_VIEWPORT_COARSE_SAMPLE_ORDER_STATE_CREATE_INFO_NV

sampleOrderType must be a valid CoarseSampleOrderTypeNV value

If customSampleOrderCount is not 0, pCustomSampleOrders must be a valid pointer to an array of customSampleOrderCount valid CoarseSampleOrderCustomNV structures

See Also

CoarseSampleOrderCustomNV, CoarseSampleOrderTypeNV, StructureType

Constructors

PipelineViewportCoarseSampleOrderStateCreateInfoNV

Fields

sampleOrderType :: CoarseSampleOrderTypeNV

sampleOrderType specifies the mechanism used to order coverage samples in fragments larger than one pixel.

customSampleOrders :: Vector CoarseSampleOrderCustomNV

pCustomSampleOrders is a pointer to an array of customSampleOrderCount CoarseSampleOrderCustomNV structures, each of which specifies the coverage sample order for a single combination of fragment area and coverage sample count.

newtype ShadingRatePaletteEntryNV

VkShadingRatePaletteEntryNV - Shading rate image palette entry types

Description

The following table indicates the width and height (in pixels) of each fragment generated using the indicated shading rate, as well as the maximum number of fragment shader invocations launched for each fragment. When processing regions of a primitive that have a shading rate of SHADING_RATE_PALETTE_ENTRY_NO_INVOCATIONS_NV, no fragments will be generated in that region.

Shading Rate 	 Width 	 Height 	 Invocations

SHADING_RATE_PALETTE_ENTRY_NO_INVOCATIONS_NV 	 0 	 0 	 0

SHADING_RATE_PALETTE_ENTRY_16_INVOCATIONS_PER_PIXEL_NV 	 1 	 1 	 16

SHADING_RATE_PALETTE_ENTRY_8_INVOCATIONS_PER_PIXEL_NV 	 1 	 1 	 8

SHADING_RATE_PALETTE_ENTRY_4_INVOCATIONS_PER_PIXEL_NV 	 1 	 1 	 4

SHADING_RATE_PALETTE_ENTRY_2_INVOCATIONS_PER_PIXEL_NV 	 1 	 1 	 2

SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_PIXEL_NV 	 1 	 1 	 1

SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_2X1_PIXELS_NV 	 2 	 1 	 1

SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_1X2_PIXELS_NV 	 1 	 2 	 1

SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_2X2_PIXELS_NV 	 2 	 2 	 1

SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_4X2_PIXELS_NV 	 4 	 2 	 1

SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_2X4_PIXELS_NV 	 2 	 4 	 1

SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_4X4_PIXELS_NV 	 4 	 4 	 1

See Also

CoarseSampleOrderCustomNV, ShadingRatePaletteNV

Constructors

ShadingRatePaletteEntryNV Int32

Bundled Patterns

pattern SHADING_RATE_PALETTE_ENTRY_NO_INVOCATIONS_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_16_INVOCATIONS_PER_PIXEL_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_8_INVOCATIONS_PER_PIXEL_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_4_INVOCATIONS_PER_PIXEL_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_2_INVOCATIONS_PER_PIXEL_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_PIXEL_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_2X1_PIXELS_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_1X2_PIXELS_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_2X2_PIXELS_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_4X2_PIXELS_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_2X4_PIXELS_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_4X4_PIXELS_NV :: ShadingRatePaletteEntryNV

newtype CoarseSampleOrderTypeNV

VkCoarseSampleOrderTypeNV - Shading rate image sample ordering types

See Also

PipelineViewportCoarseSampleOrderStateCreateInfoNV, cmdSetCoarseSampleOrderNV

Constructors

CoarseSampleOrderTypeNV Int32

Bundled Patterns

pattern COARSE_SAMPLE_ORDER_TYPE_DEFAULT_NV :: CoarseSampleOrderTypeNV
COARSE_SAMPLE_ORDER_TYPE_DEFAULT_NV specifies that coverage samples will be ordered in an implementation-dependent manner.

pattern COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV :: CoarseSampleOrderTypeNV
COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV specifies that coverage samples will be ordered according to the array of custom orderings provided in either the pCustomSampleOrders member of PipelineViewportCoarseSampleOrderStateCreateInfoNV or the pCustomSampleOrders member of cmdSetCoarseSampleOrderNV.

pattern COARSE_SAMPLE_ORDER_TYPE_PIXEL_MAJOR_NV :: CoarseSampleOrderTypeNV
COARSE_SAMPLE_ORDER_TYPE_PIXEL_MAJOR_NV specifies that coverage samples will be ordered sequentially, sorted first by pixel coordinate (in row-major order) and then by sample index.

pattern COARSE_SAMPLE_ORDER_TYPE_SAMPLE_MAJOR_NV :: CoarseSampleOrderTypeNV
COARSE_SAMPLE_ORDER_TYPE_SAMPLE_MAJOR_NV specifies that coverage samples will be ordered sequentially, sorted first by sample index and then by pixel coordinate (in row-major order).

type NV_SHADING_RATE_IMAGE_SPEC_VERSION = 3

pattern NV_SHADING_RATE_IMAGE_SPEC_VERSION :: forall a. Integral a => a

type NV_SHADING_RATE_IMAGE_EXTENSION_NAME = "VK_NV_shading_rate_image"

pattern NV_SHADING_RATE_IMAGE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a