Vulkan.Extensions.VK_QCOM_render_pass_store_ops

Description

Name

VK_QCOM_render_pass_store_ops - device extension

VK_QCOM_render_pass_store_ops

Name String

VK_QCOM_render_pass_store_ops

Extension Type

Device extension

Registered Extension Number

302

Revision

2

Extension and Version Dependencies

Requires Vulkan 1.0

Contact

Bill Licea-Kane https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_QCOM_render_pass_store_ops:%20&body=@wwlk%20

Other Extension Metadata

Last Modified Date

2020-03-25

Contributors

Bill Licea-Kane, Qualcomm Technologies, Inc.

Description

Renderpass attachments can be read-only for the duration of a renderpass.

Examples include input attachments and depth attachments where depth tests are enabled but depth writes are not enabled.

In such cases, there can be no contents generated for an attachment within the render area.

This extension adds a new AttachmentStoreOp ATTACHMENT_STORE_OP_NONE_QCOM which specifies that the contents within the render area may not be written to memory, but that the prior contents of the attachment in memory are preserved. However, if any contents were generated within the render area during rendering, the contents of the attachment will be undefined inside the render area.

Note

The AttachmentStoreOp ATTACHMENT_STORE_OP_STORE may force an implementation to assume that the attachment was written and force an implementation to flush data to memory or to a higher level cache. The AttachmentStoreOp ATTACHMENT_STORE_OP_NONE_QCOM may allow an implementation to assume that the attachment was not written and allow an implementation to avoid such a flush..

New Enum Constants

QCOM_render_pass_store_ops_EXTENSION_NAME

QCOM_render_pass_store_ops_SPEC_VERSION

Extending AttachmentStoreOp:

ATTACHMENT_STORE_OP_NONE_QCOM

Version History

Revision 1, 2019-12-20 (wwlk)

Initial version

Revision 2, 2020-03-25 (wwlk)

Minor renaming

See Also

No cross-references are available

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

type QCOM_render_pass_store_ops_SPEC_VERSION = 2

pattern QCOM_render_pass_store_ops_SPEC_VERSION :: forall a. Integral a => a

type QCOM_render_pass_store_ops_EXTENSION_NAME = "VK_QCOM_render_pass_store_ops"

pattern QCOM_render_pass_store_ops_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a