Vulkan.Extensions.VK_QCOM_render_pass_transform

Description

Name

VK_QCOM_render_pass_transform - device extension

VK_QCOM_render_pass_transform

Name String

VK_QCOM_render_pass_transform

Extension Type

Device extension

Registered Extension Number

283

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_swapchain

Requires VK_KHR_surface

Contact

Jeff Leger https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_QCOM_render_pass_transform:%20&body=@jackohound%20

Other Extension Metadata

Last Modified Date

2020-10-15

Interactions and External Dependencies

This extension requires VK_KHR_swapchain

This extension interacts with VK_EXT_fragment_density_map

Contributors

Jeff Leger, Qualcomm Technologies, Inc.

Brandon Light, Qualcomm Technologies, Inc.

Description

This extension provides a mechanism for applications to enable driver support for render pass transform.

Mobile devices can be rotated and mobile applications need to render properly when a device is held in a landscape or portrait orientation. When the current orientation differs from the device’s native orientation, a rotation is required so that the "up" direction of the rendered scene matches the current orientation.

If the Display Processing Unit (DPU) doesnt natively support rotation, the Vulkan presentation engine can handle this rotation in a separate composition pass. Alternatively, the application can render frames "pre-rotated" to avoid this extra pass. The latter is preferred to reduce power consumption and achieve the best performance because it avoids tasking the GPU with extra work to perform the copy/rotate operation.

Unlike OpenGL ES, the burden of pre-rotation in Vulkan falls on the application. To implement pre-rotation, applications render into swapchain images matching the device native aspect ratio of the display and "pre-rotate" the rendering content to match the device’s current orientation. The burden is more than adjusting the Model View Projection (MVP) matrix in the vertex shader to account for rotation and aspect ratio. The coordinate systems of scissors, viewports, derivatives and several shader built-ins may need to be adapted to produce the correct result.

It is difficult for some game engines to manage this burden; many chose to simply accept the performance/power overhead of performing rotation in the presentation engine.

This extension allows applications to achieve the performance benefits of pre-rotated rendering by moving much of the above-mentioned burden to the graphics driver. The following is unchanged with this extension:

Applications create a swapchain matching the native orientation of the display. Applications must also set the SwapchainCreateInfoKHR::preTransform equal to the currentTransform as returned by getPhysicalDeviceSurfaceCapabilitiesKHR.

The following is changed with this extension:

At cmdBeginRenderPass, the application provides extension struct RenderPassTransformBeginInfoQCOM specifying the render pass transform parameters.

At beginCommandBuffer for secondary command buffers, the application provides extension struct CommandBufferInheritanceRenderPassTransformInfoQCOM specifying the render pass transform parameters.

The renderArea, viewPorts and scissors are all provided in the current (non-rotated) coordinate system. The implementation will transform those into the native (rotated) coordinate system.

The implementation is responsible for transforming shader built-ins (FragCoord, PointCoord, SamplePosition, interpolateAt(), dFdx, dFdy, fWidth) into the rotated coordinate system.

The implementation is responsible for transforming position to the rotated coordinate system.

New Structures

Extending CommandBufferInheritanceInfo:

CommandBufferInheritanceRenderPassTransformInfoQCOM

Extending RenderPassBeginInfo:

RenderPassTransformBeginInfoQCOM

New Enum Constants

QCOM_RENDER_PASS_TRANSFORM_EXTENSION_NAME

QCOM_RENDER_PASS_TRANSFORM_SPEC_VERSION

Extending RenderPassCreateFlagBits:

RENDER_PASS_CREATE_TRANSFORM_BIT_QCOM

Extending StructureType:

STRUCTURE_TYPE_COMMAND_BUFFER_INHERITANCE_RENDER_PASS_TRANSFORM_INFO_QCOM

STRUCTURE_TYPE_RENDER_PASS_TRANSFORM_BEGIN_INFO_QCOM

Issues

1) Some early Adreno drivers (October 2019 through March 2020) advertised support for this extension but expected VK_STRUCTURE_TYPE values different from those in the vukan headers. To cover all Adreno devices on the market, applications need to detect the driver version and use the appropriate VK_STRUCTURE_TYPE values from the table below.

The driver version reported in VkPhysicalDeviceProperties.driverVersion is a uint32_t type. You can decode the uint32_t value into a major.minor.patch version as shown below:

uint32_t  major = ((driverVersion) >> 22);
uint32_t  minor = ((driverVersion) >> 12) & 0x3ff);
uint32_t  patch = ((driverVersion) & 0xfff);

If the Adreno major.minor.patch version is greater than or equal to to 512.469.0, then simply use the VK_STRUCTURE_TYPE values as defined in vulkan_core.h. If the version is less than or equal to to 512.468.0, then use the alternate values for the two VK_STRUCTURE_TYPEs in the table below.

Adreno Driver Version

512.468.0 and earlier 	 512.469.0 and later

VK_STRUCTURE_TYPE_ RENDER_PASS_TRANSFORM_BEGIN_INFO_QCOM 	 1000282000 	 1000282001

VK_STRUCTURE_TYPE_ COMMAND_BUFFER_INHERITANCE_RENDER_PASS_TRANSFORM_INFO_QCOM 	 1000282001 	 1000282000

Adreno Driver Requirements

2) Should the extension support only rotations (e.g. 90, 180, 270-degrees), or also mirror transforms (e.g. vertical flips)? Mobile use-cases only require rotation. Other display systems such as projectors might require a flipped transform.

RESOLVED: In this version of the extension, the functionality is restricted to 90, 180, and 270-degree rotations to address mobile use-cases.

3) How does this extension interact with VK_EXT_fragment_density_map?

RESOLVED Some implementations may not be able to support a render pass that enables both renderpass transform and fragment density maps. For simplicity, this extension disallows enabling both features within a single render pass.

4) What should this extension be named?

We considered names such as "rotated_rendering", "pre_rotation" and others. Since the functionality is limited to a render pass, it seemed the name should include "render_pass". While the current extension is limited to rotations, it could be extended to other transforms (like mirror) in the future.

RESOLVED The name "render_pass_transform" seems like the most accurate description of the introduced functionality.

Version History

Revision 1, 2020-02-05 (Jeff Leger)

See Also

CommandBufferInheritanceRenderPassTransformInfoQCOM, RenderPassTransformBeginInfoQCOM

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data RenderPassTransformBeginInfoQCOM

VkRenderPassTransformBeginInfoQCOM - Structure describing transform parameters of a render pass instance

Valid Usage

transform must be SURFACE_TRANSFORM_IDENTITY_BIT_KHR, SURFACE_TRANSFORM_ROTATE_90_BIT_KHR, SURFACE_TRANSFORM_ROTATE_180_BIT_KHR, or SURFACE_TRANSFORM_ROTATE_270_BIT_KHR

The renderpass must have been created with RenderPassCreateInfo::flags containing RENDER_PASS_CREATE_TRANSFORM_BIT_QCOM

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_RENDER_PASS_TRANSFORM_BEGIN_INFO_QCOM

See Also

StructureType, SurfaceTransformFlagBitsKHR

Constructors

RenderPassTransformBeginInfoQCOM

Fields

transform :: SurfaceTransformFlagBitsKHR

transform is a SurfaceTransformFlagBitsKHR value describing the transform to be applied to rasterization.

data CommandBufferInheritanceRenderPassTransformInfoQCOM

VkCommandBufferInheritanceRenderPassTransformInfoQCOM - Structure describing transformed render pass parameters command buffer

Description

When the secondary is recorded to execute within a render pass instance using cmdExecuteCommands, the render pass transform parameters of the secondary command buffer must be consistent with the render pass transform parameters specified for the render pass instance. In particular, the transform and renderArea for command buffer must be identical to the transform and renderArea of the render pass instance.

Valid Usage (Implicit)

See Also

Rect2D, StructureType, SurfaceTransformFlagBitsKHR

Constructors

CommandBufferInheritanceRenderPassTransformInfoQCOM

Fields

transform :: SurfaceTransformFlagBitsKHR

transform is a SurfaceTransformFlagBitsKHR value describing the transform to be applied to the render pass.

transform must be SURFACE_TRANSFORM_IDENTITY_BIT_KHR, SURFACE_TRANSFORM_ROTATE_90_BIT_KHR, SURFACE_TRANSFORM_ROTATE_180_BIT_KHR, or SURFACE_TRANSFORM_ROTATE_270_BIT_KHR

renderArea :: Rect2D

renderArea is the render area that is affected by the command buffer.

type QCOM_RENDER_PASS_TRANSFORM_SPEC_VERSION = 1

pattern QCOM_RENDER_PASS_TRANSFORM_SPEC_VERSION :: forall a. Integral a => a

type QCOM_RENDER_PASS_TRANSFORM_EXTENSION_NAME = "VK_QCOM_render_pass_transform"

pattern QCOM_RENDER_PASS_TRANSFORM_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype SurfaceTransformFlagBitsKHR

VkSurfaceTransformFlagBitsKHR - presentation transforms supported on a device

See Also

CommandBufferInheritanceRenderPassTransformInfoQCOM, CopyCommandTransformInfoQCOM, DisplaySurfaceCreateInfoKHR, RenderPassTransformBeginInfoQCOM, SurfaceCapabilities2EXT, SurfaceCapabilitiesKHR, SurfaceTransformFlagsKHR, SwapchainCreateInfoKHR

Constructors

SurfaceTransformFlagBitsKHR Flags

Bundled Patterns

pattern SURFACE_TRANSFORM_IDENTITY_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_IDENTITY_BIT_KHR specifies that image content is presented without being transformed.

pattern SURFACE_TRANSFORM_ROTATE_90_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_ROTATE_90_BIT_KHR specifies that image content is rotated 90 degrees clockwise.

pattern SURFACE_TRANSFORM_ROTATE_180_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_ROTATE_180_BIT_KHR specifies that image content is rotated 180 degrees clockwise.

pattern SURFACE_TRANSFORM_ROTATE_270_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_ROTATE_270_BIT_KHR specifies that image content is rotated 270 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_BIT_KHR specifies that image content is mirrored horizontally.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_90_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_90_BIT_KHR specifies that image content is mirrored horizontally, then rotated 90 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_180_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_180_BIT_KHR specifies that image content is mirrored horizontally, then rotated 180 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_270_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_270_BIT_KHR specifies that image content is mirrored horizontally, then rotated 270 degrees clockwise.

pattern SURFACE_TRANSFORM_INHERIT_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_INHERIT_BIT_KHR specifies that the presentation transform is not specified, and is instead determined by platform-specific considerations and mechanisms outside Vulkan.

type SurfaceTransformFlagsKHR = SurfaceTransformFlagBitsKHR