Vulkan.Extensions.VK_QCOM_rotated_copy_commands

Description

Name

VK_QCOM_rotated_copy_commands - device extension

VK_QCOM_rotated_copy_commands

Name String

VK_QCOM_rotated_copy_commands

Extension Type

Device extension

Registered Extension Number

334

Revision

0

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_swapchain

Requires VK_KHR_copy_commands2

Contact

Jeff Leger https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_QCOM_rotated_copy_commands:%20&body=@jackohound%20

Other Extension Metadata

Last Modified Date

2020-09-18

Interactions and External Dependencies

None

Contributors

Jeff Leger, Qualcomm Technologies, Inc.

Description

This extension extends adds an optional rotation transform to copy commands cmdBlitImage2KHR, cmdCopyImageToBuffer2KHR and cmdCopyBufferToImage2KHR. When copying between two resources, where one resource contains rotated content and the other does not, a rotated copy may be desired. This extension may be used in combination with VK_QCOM_render_pass_transform which adds rotated renderpasses.

This extension adds an extension structure to the following commands: vkCmdBlitImage2KHR, vkCmdCopyImageToBuffer2KHR and vkCmdCopyBufferToImage2KHR

Issues

1) What is an appropriate name for the added extension structure? The style guide says "Structures which extend other structures through the pNext chain should reflect the name of the base structure they extend.", but in this case a single extension structure is used to extend three base structures (vkCmdBlitImage2KHR, vkCmdCopyImageToBuffer2KHR and vkCmdCopyBufferToImage2KHR). Creating three identical structures with unique names seemed undesirable. RESOLVED: Deviate from the style guide for extension structure naming.

2) Should this extension add a rotation capability to vkCmdCopyImage2KHR?

RESOLVED: No. Use of rotated vkCmdBlitImage2KHR can fully address this use-case.

3) Should this extension add a rotation capability to vkCmdResolveImage2KHR?

RESOLVED No. Use of vkCmdResolveImage2KHR is very slow and extremely bandwidth intensive on Qualcomm’s GPU architecture and use of pResolveAttachments in vkRenderPass is the strongly preferred approach. Therefore, we choose not to introduce a rotation capability to vkCmdResolveImage2KHR.

New Structures

Extending BufferImageCopy2KHR, ImageBlit2KHR:

CopyCommandTransformInfoQCOM

New Enum Constants

QCOM_rotated_copy_commands_EXTENSION_NAME

QCOM_rotated_copy_commands_SPEC_VERSION

Extending StructureType:

STRUCTURE_TYPE_COPY_COMMAND_TRANSFORM_INFO_QCOM

Version History

Revision 1, 2020-09-19 (Jeff Leger)

See Also

CopyCommandTransformInfoQCOM

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data CopyCommandTransformInfoQCOM

VkCopyCommandTransformInfoQCOM - Structure describing transform parameters of rotated copy command

Valid Usage (Implicit)

See Also

StructureType, SurfaceTransformFlagBitsKHR

Constructors

CopyCommandTransformInfoQCOM

Fields

transform :: SurfaceTransformFlagBitsKHR

transform is a SurfaceTransformFlagBitsKHR value describing the transform to be applied.

transform must be SURFACE_TRANSFORM_IDENTITY_BIT_KHR, SURFACE_TRANSFORM_ROTATE_90_BIT_KHR, SURFACE_TRANSFORM_ROTATE_180_BIT_KHR, or SURFACE_TRANSFORM_ROTATE_270_BIT_KHR

type QCOM_rotated_copy_commands_SPEC_VERSION = 0

pattern QCOM_rotated_copy_commands_SPEC_VERSION :: forall a. Integral a => a

type QCOM_rotated_copy_commands_EXTENSION_NAME = "VK_QCOM_rotated_copy_commands"

pattern QCOM_rotated_copy_commands_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype SurfaceTransformFlagBitsKHR

VkSurfaceTransformFlagBitsKHR - presentation transforms supported on a device

See Also

CommandBufferInheritanceRenderPassTransformInfoQCOM, CopyCommandTransformInfoQCOM, DisplaySurfaceCreateInfoKHR, RenderPassTransformBeginInfoQCOM, SurfaceCapabilities2EXT, SurfaceCapabilitiesKHR, SurfaceTransformFlagsKHR, SwapchainCreateInfoKHR

Constructors

SurfaceTransformFlagBitsKHR Flags

Bundled Patterns

pattern SURFACE_TRANSFORM_IDENTITY_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_IDENTITY_BIT_KHR specifies that image content is presented without being transformed.

pattern SURFACE_TRANSFORM_ROTATE_90_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_ROTATE_90_BIT_KHR specifies that image content is rotated 90 degrees clockwise.

pattern SURFACE_TRANSFORM_ROTATE_180_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_ROTATE_180_BIT_KHR specifies that image content is rotated 180 degrees clockwise.

pattern SURFACE_TRANSFORM_ROTATE_270_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_ROTATE_270_BIT_KHR specifies that image content is rotated 270 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_BIT_KHR specifies that image content is mirrored horizontally.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_90_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_90_BIT_KHR specifies that image content is mirrored horizontally, then rotated 90 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_180_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_180_BIT_KHR specifies that image content is mirrored horizontally, then rotated 180 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_270_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_270_BIT_KHR specifies that image content is mirrored horizontally, then rotated 270 degrees clockwise.

pattern SURFACE_TRANSFORM_INHERIT_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_INHERIT_BIT_KHR specifies that the presentation transform is not specified, and is instead determined by platform-specific considerations and mechanisms outside Vulkan.

type SurfaceTransformFlagsKHR = SurfaceTransformFlagBitsKHR