Vulkan.Extensions.VK_VALVE_mutable_descriptor_type

Description

Name

VK_VALVE_mutable_descriptor_type - device extension

VK_VALVE_mutable_descriptor_type

Name String

VK_VALVE_mutable_descriptor_type

Extension Type

Device extension

Registered Extension Number

352

Revision

1

Extension and Version Dependencies

Requires Vulkan 1.0

Requires VK_KHR_maintenance3

Special Use

D3D support

Contact

Joshua Ashton https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_VALVE_mutable_descriptor_type:%20&body=@Joshua-Ashton%20

Hans-Kristian Arntzen https://github.com/KhronosGroup/Vulkan-Docs/issues/new?title=VK_VALVE_mutable_descriptor_type:%20&body=@HansKristian-Work%20

Other Extension Metadata

Last Modified Date

2020-12-02

IP Status

No known IP claims.

Contributors

Joshua Ashton, Valve

Hans-Kristian Arntzen, Valve

Description

This extension allows applications to reduce descriptor memory footprint by allowing a descriptor to be able to mutate to a given list of descriptor types depending on which descriptor types are written into, or copied into a descriptor set.

The main use case this extension intends to address is descriptor indexing with DESCRIPTOR_BINDING_VARIABLE_DESCRIPTOR_COUNT_BIT where the descriptor types are completely generic, as this means applications can allocate one large descriptor set, rather than having one large descriptor set per descriptor type, which significantly bloats descriptor memory usage and causes performance issues.

This extension also adds a mechanism to declare that a descriptor pool, and therefore the descriptor sets that are allocated from it, reside only in host memory; as such these descriptors can only be updated/copied, but not bound.

These features together allow much more efficient emulation of the raw D3D12 binding model. This extension is primarily intended to be useful for API layering efforts.

New Structures

MutableDescriptorTypeListVALVE

Extending DescriptorSetLayoutCreateInfo, DescriptorPoolCreateInfo:

MutableDescriptorTypeCreateInfoVALVE

Extending PhysicalDeviceFeatures2, DeviceCreateInfo:

PhysicalDeviceMutableDescriptorTypeFeaturesVALVE

New Enum Constants

VALVE_MUTABLE_DESCRIPTOR_TYPE_EXTENSION_NAME

VALVE_MUTABLE_DESCRIPTOR_TYPE_SPEC_VERSION

Extending DescriptorPoolCreateFlagBits:

DESCRIPTOR_POOL_CREATE_HOST_ONLY_BIT_VALVE

Extending DescriptorSetLayoutCreateFlagBits:

DESCRIPTOR_SET_LAYOUT_CREATE_HOST_ONLY_POOL_BIT_VALVE

Extending DescriptorType:

DESCRIPTOR_TYPE_MUTABLE_VALVE

Extending StructureType:

STRUCTURE_TYPE_MUTABLE_DESCRIPTOR_TYPE_CREATE_INFO_VALVE

STRUCTURE_TYPE_PHYSICAL_DEVICE_MUTABLE_DESCRIPTOR_TYPE_FEATURES_VALVE

Version History

Revision 1, 2020-12-01 (Joshua Ashton, Hans-Kristian Arntzen)

Initial specification, squashed from public draft.

See Also

MutableDescriptorTypeCreateInfoVALVE, MutableDescriptorTypeListVALVE, PhysicalDeviceMutableDescriptorTypeFeaturesVALVE

Document Notes

For more information, see the Vulkan Specification

This page is a generated document. Fixes and changes should be made to the generator scripts, not directly.

Documentation

data PhysicalDeviceMutableDescriptorTypeFeaturesVALVE

VkPhysicalDeviceMutableDescriptorTypeFeaturesVALVE - Structure describing whether the mutable descriptor type is supported

Members

The members of the PhysicalDeviceMutableDescriptorTypeFeaturesVALVE structure describe the following features:

Description

mutableDescriptorType indicates that the implementation must support using the DescriptorType of DESCRIPTOR_TYPE_MUTABLE_VALVE with at least the following descriptor types, where any combination of the types must be supported:

DESCRIPTOR_TYPE_SAMPLED_IMAGE

DESCRIPTOR_TYPE_STORAGE_IMAGE

DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER

DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER

DESCRIPTOR_TYPE_UNIFORM_BUFFER

DESCRIPTOR_TYPE_STORAGE_BUFFER

Additionally, mutableDescriptorType indicates that:

Non-uniform descriptor indexing must be supported if all descriptor types in a MutableDescriptorTypeListVALVE for DESCRIPTOR_TYPE_MUTABLE_VALVE have the corresponding non-uniform indexing features enabled in PhysicalDeviceDescriptorIndexingFeatures.

DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT with descriptorType of DESCRIPTOR_TYPE_MUTABLE_VALVE relaxes the list of required descriptor types to the descriptor types which have the corresponding update-after-bind feature enabled in PhysicalDeviceDescriptorIndexingFeatures.

Dynamically uniform descriptor indexing must be supported if all descriptor types in a MutableDescriptorTypeListVALVE for DESCRIPTOR_TYPE_MUTABLE_VALVE have the corresponding dynamic indexing features enabled.

DESCRIPTOR_SET_LAYOUT_CREATE_HOST_ONLY_POOL_BIT_VALVE must be supported.

DESCRIPTOR_POOL_CREATE_HOST_ONLY_BIT_VALVE must be supported.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PHYSICAL_DEVICE_MUTABLE_DESCRIPTOR_TYPE_FEATURES_VALVE

See Also

Bool32, StructureType

Constructors

PhysicalDeviceMutableDescriptorTypeFeaturesVALVE

Fields

mutableDescriptorType :: Bool

data MutableDescriptorTypeListVALVE

VkMutableDescriptorTypeListVALVE - Structure describing descriptor types that a given descriptor may mutate to

Valid Usage

descriptorTypeCount must not be 0 if the corresponding binding is of DESCRIPTOR_TYPE_MUTABLE_VALVE

pDescriptorTypes must be a valid pointer to an array of descriptorTypeCount valid, unique DescriptorType values if the given binding is of DESCRIPTOR_TYPE_MUTABLE_VALVE type

descriptorTypeCount must be 0 if the corresponding binding is not of DESCRIPTOR_TYPE_MUTABLE_VALVE

pDescriptorTypes must not contain DESCRIPTOR_TYPE_MUTABLE_VALVE

pDescriptorTypes must not contain DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC

pDescriptorTypes must not contain DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC

pDescriptorTypes must not contain DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT

Valid Usage (Implicit)

If descriptorTypeCount is not 0, pDescriptorTypes must be a valid pointer to an array of descriptorTypeCount valid DescriptorType values

See Also

DescriptorType, MutableDescriptorTypeCreateInfoVALVE

Constructors

MutableDescriptorTypeListVALVE

Fields

descriptorTypes :: Vector DescriptorType

pDescriptorTypes is NULL or a pointer to an array of descriptorTypeCount DescriptorType values which define which descriptor types a given binding may mutate to.

data MutableDescriptorTypeCreateInfoVALVE

VkMutableDescriptorTypeCreateInfoVALVE - Structure describing the list of possible active descriptor types for mutable type descriptors

Description

If mutableDescriptorTypeListCount is zero or if this structure is not included in the pNext chain, the MutableDescriptorTypeListVALVE for each element is considered to be zero or NULL for each member. Otherwise, the descriptor set layout binding at DescriptorSetLayoutCreateInfo::pBindings[i] uses the descriptor type lists in MutableDescriptorTypeCreateInfoVALVE::pMutableDescriptorTypeLists[i].

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_MUTABLE_DESCRIPTOR_TYPE_CREATE_INFO_VALVE

If mutableDescriptorTypeListCount is not 0, pMutableDescriptorTypeLists must be a valid pointer to an array of mutableDescriptorTypeListCount valid MutableDescriptorTypeListVALVE structures

See Also

MutableDescriptorTypeListVALVE, StructureType

Constructors

MutableDescriptorTypeCreateInfoVALVE

Fields

mutableDescriptorTypeLists :: Vector MutableDescriptorTypeListVALVE

pMutableDescriptorTypeLists is a pointer to an array of MutableDescriptorTypeListVALVE structures.

type VALVE_MUTABLE_DESCRIPTOR_TYPE_SPEC_VERSION = 1

pattern VALVE_MUTABLE_DESCRIPTOR_TYPE_SPEC_VERSION :: forall a. Integral a => a

type VALVE_MUTABLE_DESCRIPTOR_TYPE_EXTENSION_NAME = "VK_VALVE_mutable_descriptor_type"

pattern VALVE_MUTABLE_DESCRIPTOR_TYPE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a