Vulkan.Internal.Utils

Documentation

enumReadPrec

Arguments

:: Read i

=> String
The common constructor prefix

-> [(a, String)]
The table of values to constructor suffixes

-> String
The newtype constructor name

-> (i -> a)
The newtype constructor

-> ReadPrec a

The common bits of enumeration and bitmask read instances

enumShowsPrec

Arguments

:: Eq a

=> String
The common constructor prefix

-> [(a, String)]
A table of values to constructor suffixes

-> String
The newtype constructor name

-> (a -> i)
Unpack the newtype

-> (i -> ShowS)
Show the underlying value

-> Int

-> a

-> ShowS

The common bits of enumeration and bitmask show instances