Vulkan.Requirement

Vulkan requirements

data InstanceRequirement where

A requirement on a Vulkan Instance.

Constructors

RequireInstanceVersion
Require a minimum Vulkan instance version.

Fields

:: { version :: Word32

} -> InstanceRequirement

RequireInstanceLayer
Require a Vulkan layer.

Fields

:: { instanceLayerName :: ByteString

, instanceLayerMinVersion :: Word32

The implementationVersion of the layer must meet or exceed this version

} -> InstanceRequirement

RequireInstanceExtension
Require a Vulkan instance extension.

Fields

:: { instanceExtensionLayerName :: Maybe ByteString

, instanceExtensionName :: ByteString

, instanceExtensionMinVersion :: Word32

} -> InstanceRequirement

data DeviceRequirement where

Constructors

RequireDeviceVersion
Require a minimum device version.

Fields

:: { version :: Word32

} -> DeviceRequirement

RequireDeviceFeature
Require a Vulkan physical device feature.

Fields

:: forall struct. KnownFeatureStruct struct

=> { featureName :: ByteString

, checkFeature :: struct -> Bool

, enableFeature :: struct -> struct

} -> DeviceRequirement

RequireDeviceProperty
Require a Vulkan physical device property.

Fields

:: forall struct. KnownPropertyStruct struct

=> { propertyName :: ByteString

, checkProperty :: struct -> Bool

} -> DeviceRequirement

RequireDeviceExtension
Require a Vulkan device extension.

Fields

:: { deviceExtensionLayerName :: Maybe ByteString

, deviceExtensionName :: ByteString

, deviceExtensionMinVersion :: Word32

} -> DeviceRequirement

Utility functionality for handling structure chains

class Typeable feat => KnownFeatureStruct feat where

A Vulkan structure that can appear in PhysicalDeviceFeatures2.

Methods

sFeatureStruct :: SFeatureStruct feat

class Typeable prop => KnownPropertyStruct prop where

A Vulkan structure that can appear in PhysicalDeviceProperties2.

Methods

sPropertyStruct :: SPropertyStruct prop

data SFeatureStruct feat where

Singleton for a Vulkan structure that can appear in PhysicalDeviceFeatures2.

It is either PhysicalDeviceFeatures, or it Extends PhysicalDeviceFeatures2.

Constructors

BasicFeatureStruct :: SFeatureStruct PhysicalDeviceFeatures

ExtendedFeatureStruct :: (Show feat, Extends PhysicalDeviceFeatures2 feat, Extends DeviceCreateInfo feat, Zero feat, FromCStruct feat, ToCStruct feat) => SFeatureStruct feat

data SPropertyStruct prop where

Singleton for a Vulkan structure that can appear in PhysicalDeviceProperties2.

It is either PhysicalDeviceProperties, or it Extends PhysicalDeviceProperties2.

Constructors

BasicPropertyStruct :: SPropertyStruct PhysicalDeviceProperties

ExtendedPropertyStruct :: (Typeable prop, Extends PhysicalDeviceProperties2 prop, FromCStruct prop, ToCStruct prop) => SPropertyStruct prop