Vulkan.Zero

Documentation

class Zero a where

A class for initializing things with all zero data

Any instance should satisfy the following law:

new zero = calloc or with zero = withZeroCStruct

i.e. Marshaling zero to memory yeilds only zero-valued bytes, except for structs which require a "type" tag

Methods

zero :: a