Graphics.Vulkan.Core10.Buffer

Documentation

type BufferCreateFlagBits = VkBufferCreateFlagBits

VkBufferCreateFlagBits - Bitmask specifying additional parameters of a buffer

Description

See Sparse Resource Features and Physical Device Features for details of the sparse memory features supported on a device.

See Also

VkBufferCreateFlags

pattern BUFFER_CREATE_SPARSE_BINDING_BIT :: a ~ BufferCreateFlagBits => a

VK_BUFFER_CREATE_SPARSE_BINDING_BIT specifies that the buffer will be backed using sparse memory binding.

pattern BUFFER_CREATE_SPARSE_RESIDENCY_BIT :: a ~ BufferCreateFlagBits => a

VK_BUFFER_CREATE_SPARSE_RESIDENCY_BIT specifies that the buffer can be partially backed using sparse memory binding. Buffers created with this flag must also be created with the VK_BUFFER_CREATE_SPARSE_BINDING_BIT flag.

pattern BUFFER_CREATE_SPARSE_ALIASED_BIT :: a ~ BufferCreateFlagBits => a

VK_BUFFER_CREATE_SPARSE_ALIASED_BIT specifies that the buffer will be backed using sparse memory binding with memory ranges that might also simultaneously be backing another buffer (or another portion of the same buffer). Buffers created with this flag must also be created with the VK_BUFFER_CREATE_SPARSE_BINDING_BIT flag.

pattern BUFFER_CREATE_PROTECTED_BIT :: a ~ BufferCreateFlagBits => a

pattern BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT_EXT :: a ~ BufferCreateFlagBits => a

type BufferCreateFlags = BufferCreateFlagBits

VkBufferCreateFlags - Bitmask of VkBufferCreateFlagBits

Description

VkBufferCreateFlags is a bitmask type for setting a mask of zero or more VkBufferCreateFlagBits.

See Also

VkBufferCreateFlagBits, VkBufferCreateInfo, VkPhysicalDeviceExternalBufferInfo

withCStructBufferCreateInfo :: BufferCreateInfo -> (VkBufferCreateInfo -> IO a) -> IO a

A function to temporarily allocate memory for a VkBufferCreateInfo and marshal a BufferCreateInfo into it. The VkBufferCreateInfo is only valid inside the provided computation and must not be returned out of it.

fromCStructBufferCreateInfo :: VkBufferCreateInfo -> IO BufferCreateInfo

A function to read a VkBufferCreateInfo and all additional structures in the pointer chain into a BufferCreateInfo.

data BufferCreateInfo

VkBufferCreateInfo - Structure specifying the parameters of a newly created buffer object

Valid Usage

size must be greater than 0

If sharingMode is VK_SHARING_MODE_CONCURRENT, pQueueFamilyIndices must be a valid pointer to an array of queueFamilyIndexCount uint32_t values

If sharingMode is VK_SHARING_MODE_CONCURRENT, queueFamilyIndexCount must be greater than 1

If sharingMode is VK_SHARING_MODE_CONCURRENT, each element of pQueueFamilyIndices must be unique and must be less than pQueueFamilyPropertyCount returned by vkGetPhysicalDeviceQueueFamilyProperties for the physicalDevice that was used to create device

If the sparse bindings feature is not enabled, flags must not contain VK_BUFFER_CREATE_SPARSE_BINDING_BIT

If the sparse buffer residency feature is not enabled, flags must not contain VK_BUFFER_CREATE_SPARSE_RESIDENCY_BIT

If the sparse aliased residency feature is not enabled, flags must not contain VK_BUFFER_CREATE_SPARSE_ALIASED_BIT

If flags contains VK_BUFFER_CREATE_SPARSE_RESIDENCY_BIT or VK_BUFFER_CREATE_SPARSE_ALIASED_BIT, it must also contain VK_BUFFER_CREATE_SPARSE_BINDING_BIT

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_BUFFER_CREATE_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of VkBufferDeviceAddressCreateInfoEXT, VkDedicatedAllocationBufferCreateInfoNV, or VkExternalMemoryBufferCreateInfo

Each sType member in the pNext chain must be unique

flags must be a valid combination of VkBufferCreateFlagBits values

usage must be a valid combination of VkBufferUsageFlagBits values

usage must not be 0

sharingMode must be a valid VkSharingMode value

See Also

VkBufferCreateFlags, VkBufferUsageFlags, VkDeviceSize, VkSharingMode, VkStructureType, vkCreateBuffer

Constructors

BufferCreateInfo

Fields

next :: Maybe SomeVkStruct

flags :: BufferCreateFlags

size :: DeviceSize

usage :: BufferUsageFlags

sharingMode :: SharingMode

queueFamilyIndices :: Vector Word32

type BufferUsageFlagBits = VkBufferUsageFlagBits

VkBufferUsageFlagBits - Bitmask specifying allowed usage of a buffer

See Also

VkBufferUsageFlags

pattern BUFFER_USAGE_TRANSFER_SRC_BIT :: a ~ BufferUsageFlagBits => a

VK_BUFFER_USAGE_TRANSFER_SRC_BIT specifies that the buffer can be used as the source of a /transfer command/ (see the definition of https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#synchronization-pipeline-stages-transfer).

pattern BUFFER_USAGE_TRANSFER_DST_BIT :: a ~ BufferUsageFlagBits => a

VK_BUFFER_USAGE_TRANSFER_DST_BIT specifies that the buffer can be used as the destination of a transfer command.

pattern BUFFER_USAGE_UNIFORM_TEXEL_BUFFER_BIT :: a ~ BufferUsageFlagBits => a

VK_BUFFER_USAGE_UNIFORM_TEXEL_BUFFER_BIT specifies that the buffer can be used to create a VkBufferView suitable for occupying a VkDescriptorSet slot of type VK_DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER.

pattern BUFFER_USAGE_STORAGE_TEXEL_BUFFER_BIT :: a ~ BufferUsageFlagBits => a

VK_BUFFER_USAGE_STORAGE_TEXEL_BUFFER_BIT specifies that the buffer can be used to create a VkBufferView suitable for occupying a VkDescriptorSet slot of type VK_DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER.

pattern BUFFER_USAGE_UNIFORM_BUFFER_BIT :: a ~ BufferUsageFlagBits => a

VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT specifies that the buffer can be used in a VkDescriptorBufferInfo suitable for occupying a VkDescriptorSet slot either of type VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER or VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC.

pattern BUFFER_USAGE_STORAGE_BUFFER_BIT :: a ~ BufferUsageFlagBits => a

VK_BUFFER_USAGE_STORAGE_BUFFER_BIT specifies that the buffer can be used in a VkDescriptorBufferInfo suitable for occupying a VkDescriptorSet slot either of type VK_DESCRIPTOR_TYPE_STORAGE_BUFFER or VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC.

pattern BUFFER_USAGE_INDEX_BUFFER_BIT :: a ~ BufferUsageFlagBits => a

VK_BUFFER_USAGE_INDEX_BUFFER_BIT specifies that the buffer is suitable for passing as the buffer parameter to vkCmdBindIndexBuffer.

pattern BUFFER_USAGE_VERTEX_BUFFER_BIT :: a ~ BufferUsageFlagBits => a

VK_BUFFER_USAGE_VERTEX_BUFFER_BIT specifies that the buffer is suitable for passing as an element of the pBuffers array to vkCmdBindVertexBuffers.

pattern BUFFER_USAGE_INDIRECT_BUFFER_BIT :: a ~ BufferUsageFlagBits => a

VK_BUFFER_USAGE_INDIRECT_BUFFER_BIT specifies that the buffer is suitable for passing as the buffer parameter to vkCmdDrawIndirect, vkCmdDrawIndexedIndirect, or vkCmdDispatchIndirect.

pattern BUFFER_USAGE_RESERVED_15_BIT_KHR :: a ~ BufferUsageFlagBits => a

pattern BUFFER_USAGE_RESERVED_16_BIT_KHR :: a ~ BufferUsageFlagBits => a

pattern BUFFER_USAGE_RESERVED_13_BIT_KHR :: a ~ BufferUsageFlagBits => a

pattern BUFFER_USAGE_RESERVED_14_BIT_KHR :: a ~ BufferUsageFlagBits => a

pattern BUFFER_USAGE_TRANSFORM_FEEDBACK_BUFFER_BIT_EXT :: a ~ BufferUsageFlagBits => a

pattern BUFFER_USAGE_TRANSFORM_FEEDBACK_COUNTER_BUFFER_BIT_EXT :: a ~ BufferUsageFlagBits => a

pattern BUFFER_USAGE_CONDITIONAL_RENDERING_BIT_EXT :: a ~ BufferUsageFlagBits => a

pattern BUFFER_USAGE_RAY_TRACING_BIT_NV :: a ~ BufferUsageFlagBits => a

pattern BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT_EXT :: a ~ BufferUsageFlagBits => a

type BufferUsageFlags = BufferUsageFlagBits

VkBufferUsageFlags - Bitmask of VkBufferUsageFlagBits

Description

VkBufferUsageFlags is a bitmask type for setting a mask of zero or more VkBufferUsageFlagBits.

See Also

VkBufferCreateInfo, VkBufferUsageFlagBits, VkPhysicalDeviceExternalBufferInfo

type SharingMode = VkSharingMode

VkSharingMode - Buffer and image sharing modes

Description

Note

VK_SHARING_MODE_CONCURRENT may result in lower performance access to the buffer or image than VK_SHARING_MODE_EXCLUSIVE.

Ranges of buffers and image subresources of image objects created using VK_SHARING_MODE_EXCLUSIVE must only be accessed by queues in the queue family that has ownership of the resource. Upon creation, such resources are not owned by any queue family; ownership is implicitly acquired upon first use within a queue. Once a resource using VK_SHARING_MODE_EXCLUSIVE is owned by some queue family, the application must perform a queue family ownership transfer to make the memory contents of a range or image subresource accessible to a different queue family.

Note

Images still require a layout transition from VK_IMAGE_LAYOUT_UNDEFINED or VK_IMAGE_LAYOUT_PREINITIALIZED before being used on the first queue.

A queue family can take ownership of an image subresource or buffer range of a resource created with VK_SHARING_MODE_EXCLUSIVE, without an ownership transfer, in the same way as for a resource that was just created; however, taking ownership in this way has the effect that the contents of the image subresource or buffer range are undefined.

Ranges of buffers and image subresources of image objects created using VK_SHARING_MODE_CONCURRENT must only be accessed by queues from the queue families specified through the queueFamilyIndexCount and pQueueFamilyIndices members of the corresponding create info structures.

See Also

VkBufferCreateInfo, VkImageCreateInfo, VkPhysicalDeviceImageDrmFormatModifierInfoEXT, VkSwapchainCreateInfoKHR

pattern SHARING_MODE_EXCLUSIVE :: a ~ SharingMode => a

VK_SHARING_MODE_EXCLUSIVE specifies that access to any range or image subresource of the object will be exclusive to a single queue family at a time.

pattern SHARING_MODE_CONCURRENT :: a ~ SharingMode => a

VK_SHARING_MODE_CONCURRENT specifies that concurrent access to any range or image subresource of the object from multiple queue families is supported.

createBuffer :: Device -> BufferCreateInfo -> Maybe AllocationCallbacks -> IO Buffer

vkCreateBuffer - Create a new buffer object

Parameters

device is the logical device that creates the buffer object.

pCreateInfo is a pointer to an instance of the VkBufferCreateInfo structure containing parameters affecting creation of the buffer.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pBuffer points to a VkBuffer handle in which the resulting buffer object is returned.

Valid Usage

If the flags member of pCreateInfo includes VK_BUFFER_CREATE_SPARSE_BINDING_BIT, creating this VkBuffer must not cause the total required sparse memory for all currently valid sparse resources on the device to exceed VkPhysicalDeviceLimits::sparseAddressSpaceSize

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkBufferCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pBuffer must be a valid pointer to a VkBuffer handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_INVALID_DEVICE_ADDRESS_EXT

See Also

VkAllocationCallbacks, VkBuffer, VkBufferCreateInfo, VkDevice

destroyBuffer :: Device -> Buffer -> Maybe AllocationCallbacks -> IO ()

vkDestroyBuffer - Destroy a buffer object

Parameters

device is the logical device that destroys the buffer.

buffer is the buffer to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to buffer, either directly or via a VkBufferView, must have completed execution

If VkAllocationCallbacks were provided when buffer was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when buffer was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid VkDevice handle

If buffer is not VK_NULL_HANDLE, buffer must be a valid VkBuffer handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

If buffer is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to buffer must be externally synchronized

See Also

VkAllocationCallbacks, VkBuffer, VkDevice

withBuffer :: Device -> BufferCreateInfo -> Maybe AllocationCallbacks -> (Buffer -> IO a) -> IO a

A safe wrapper for createBuffer and destroyBuffer using bracket

The allocated value must not be returned from the provided computation

pattern VK_BUFFER_USAGE_RESERVED_13_BIT_KHR :: VkBufferUsageFlagBits

pattern VK_BUFFER_USAGE_RESERVED_14_BIT_KHR :: VkBufferUsageFlagBits

pattern VK_BUFFER_USAGE_RESERVED_15_BIT_KHR :: VkBufferUsageFlagBits

pattern VK_BUFFER_USAGE_RESERVED_16_BIT_KHR :: VkBufferUsageFlagBits