Graphics.Vulkan.Core10.DescriptorSet

Documentation

createDescriptorSetLayout :: PokeChain a => Device -> DescriptorSetLayoutCreateInfo a -> ("allocator" ::: Maybe AllocationCallbacks) -> IO DescriptorSetLayout

vkCreateDescriptorSetLayout - Create a new descriptor set layout

Parameters

Device is the logical device that creates the descriptor set layout.

pCreateInfo is a pointer to a DescriptorSetLayoutCreateInfo structure specifying the state of the descriptor set layout object.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pSetLayout is a pointer to a DescriptorSetLayout handle in which the resulting descriptor set layout object is returned.

Valid Usage (Implicit)

Device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid DescriptorSetLayoutCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pSetLayout must be a valid pointer to a DescriptorSetLayout handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, DescriptorSetLayout, DescriptorSetLayoutCreateInfo, Device

withDescriptorSetLayout :: PokeChain a => Device -> DescriptorSetLayoutCreateInfo a -> Maybe AllocationCallbacks -> (DescriptorSetLayout -> IO r) -> IO r

A safe wrapper for createDescriptorSetLayout and destroyDescriptorSetLayout using bracket

The allocated value must not be returned from the provided computation

destroyDescriptorSetLayout :: Device -> DescriptorSetLayout -> ("allocator" ::: Maybe AllocationCallbacks) -> IO ()

vkDestroyDescriptorSetLayout - Destroy a descriptor set layout object

Parameters

Device is the logical device that destroys the descriptor set layout.

DescriptorSetLayout is the descriptor set layout to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

If AllocationCallbacks were provided when DescriptorSetLayout was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when DescriptorSetLayout was created, pAllocator must be NULL

Valid Usage (Implicit)

Device must be a valid Device handle

If DescriptorSetLayout is not NULL_HANDLE, DescriptorSetLayout must be a valid DescriptorSetLayout handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If DescriptorSetLayout is a valid handle, it must have been created, allocated, or retrieved from Device

Host Synchronization

Host access to DescriptorSetLayout must be externally synchronized

See Also

AllocationCallbacks, DescriptorSetLayout, Device

createDescriptorPool :: PokeChain a => Device -> DescriptorPoolCreateInfo a -> ("allocator" ::: Maybe AllocationCallbacks) -> IO DescriptorPool

vkCreateDescriptorPool - Creates a descriptor pool object

Parameters

Device is the logical device that creates the descriptor pool.

pCreateInfo is a pointer to a DescriptorPoolCreateInfo structure specifying the state of the descriptor pool object.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pDescriptorPool is a pointer to a DescriptorPool handle in which the resulting descriptor pool object is returned.

Description

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

The created descriptor pool is returned in pDescriptorPool.

Valid Usage (Implicit)

Device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid DescriptorPoolCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pDescriptorPool must be a valid pointer to a DescriptorPool handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_FRAGMENTATION_EXT

See Also

AllocationCallbacks, DescriptorPool, DescriptorPoolCreateInfo, Device

withDescriptorPool :: PokeChain a => Device -> DescriptorPoolCreateInfo a -> Maybe AllocationCallbacks -> (DescriptorPool -> IO r) -> IO r

A safe wrapper for createDescriptorPool and destroyDescriptorPool using bracket

The allocated value must not be returned from the provided computation

destroyDescriptorPool :: Device -> DescriptorPool -> ("allocator" ::: Maybe AllocationCallbacks) -> IO ()

vkDestroyDescriptorPool - Destroy a descriptor pool object

Parameters

Device is the logical device that destroys the descriptor pool.

DescriptorPool is the descriptor pool to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Description

When a pool is destroyed, all descriptor sets allocated from the pool are implicitly freed and become invalid. Descriptor sets allocated from a given pool do not need to be freed before destroying that descriptor pool.

Valid Usage

All submitted commands that refer to DescriptorPool (via any allocated descriptor sets) must have completed execution

If AllocationCallbacks were provided when DescriptorPool was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when DescriptorPool was created, pAllocator must be NULL

Valid Usage (Implicit)

Device must be a valid Device handle

If DescriptorPool is not NULL_HANDLE, DescriptorPool must be a valid DescriptorPool handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If DescriptorPool is a valid handle, it must have been created, allocated, or retrieved from Device

Host Synchronization

Host access to DescriptorPool must be externally synchronized

See Also

AllocationCallbacks, DescriptorPool, Device

resetDescriptorPool :: Device -> DescriptorPool -> DescriptorPoolResetFlags -> IO ()

vkResetDescriptorPool - Resets a descriptor pool object

Parameters

Device is the logical device that owns the descriptor pool.

DescriptorPool is the descriptor pool to be reset.

Flags is reserved for future use.

Description

Resetting a descriptor pool recycles all of the resources from all of the descriptor sets allocated from the descriptor pool back to the descriptor pool, and the descriptor sets are implicitly freed.

Valid Usage

All uses of DescriptorPool (via any allocated descriptor sets) must have completed execution

Valid Usage (Implicit)

Device must be a valid Device handle

DescriptorPool must be a valid DescriptorPool handle

Flags must be 0

DescriptorPool must have been created, allocated, or retrieved from Device

Host Synchronization

Host access to DescriptorPool must be externally synchronized

Host access to any DescriptorSet objects allocated from DescriptorPool must be externally synchronized

Return Codes

Success

SUCCESS

See Also

DescriptorPool, DescriptorPoolResetFlags, Device

allocateDescriptorSets :: PokeChain a => Device -> DescriptorSetAllocateInfo a -> IO ("descriptorSets" ::: Vector DescriptorSet)

vkAllocateDescriptorSets - Allocate one or more descriptor sets

Parameters

Device is the logical device that owns the descriptor pool.

pAllocateInfo is a pointer to a DescriptorSetAllocateInfo structure describing parameters of the allocation.

pDescriptorSets is a pointer to an array of DescriptorSet handles in which the resulting descriptor set objects are returned.

Description

The allocated descriptor sets are returned in pDescriptorSets.

When a descriptor set is allocated, the initial state is largely uninitialized and all descriptors are undefined. Descriptors also become undefined if the underlying resource is destroyed. Descriptor sets containing undefined descriptors can still be bound and used, subject to the following conditions:

For descriptor set bindings created with the DESCRIPTOR_BINDING_PARTIALLY_BOUND_BIT bit set, all descriptors in that binding that are dynamically used must have been populated before the descriptor set is consumed.

For descriptor set bindings created without the DESCRIPTOR_BINDING_PARTIALLY_BOUND_BIT bit set, all descriptors in that binding that are statically used must have been populated before the descriptor set is consumed.

Descriptor bindings with descriptor type of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT can be undefined when the descriptor set is consumed; though values in that block will be undefined.

Entries that are not used by a pipeline can have undefined descriptors.

If a call to allocateDescriptorSets would cause the total number of descriptor sets allocated from the pool to exceed the value of DescriptorPoolCreateInfo::maxSets used to create pAllocateInfo->descriptorPool, then the allocation may fail due to lack of space in the descriptor pool. Similarly, the allocation may fail due to lack of space if the call to allocateDescriptorSets would cause the number of any given descriptor type to exceed the sum of all the descriptorCount members of each element of DescriptorPoolCreateInfo::pPoolSizes with a member equal to that type.

Additionally, the allocation may also fail if a call to allocateDescriptorSets would cause the total number of inline uniform block bindings allocated from the pool to exceed the value of DescriptorPoolInlineUniformBlockCreateInfoEXT::maxInlineUniformBlockBindings used to create the descriptor pool.

If the allocation fails due to no more space in the descriptor pool, and not because of system or device memory exhaustion, then ERROR_OUT_OF_POOL_MEMORY must be returned.

allocateDescriptorSets can be used to create multiple descriptor sets. If the creation of any of those descriptor sets fails, then the implementation must destroy all successfully created descriptor set objects from this command, set all entries of the pDescriptorSets array to NULL_HANDLE and return the error.

Valid Usage (Implicit)

Device must be a valid Device handle

pAllocateInfo must be a valid pointer to a valid DescriptorSetAllocateInfo structure

pDescriptorSets must be a valid pointer to an array of pAllocateInfo::descriptorSetCount DescriptorSet handles

The value referenced by pAllocateInfo::descriptorSetCount must be greater than 0

Host Synchronization

Host access to pAllocateInfo::descriptorPool must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_FRAGMENTED_POOL

ERROR_OUT_OF_POOL_MEMORY

See Also

DescriptorSet, DescriptorSetAllocateInfo, Device

withDescriptorSets :: PokeChain a => Device -> DescriptorSetAllocateInfo a -> (Vector DescriptorSet -> IO r) -> IO r

A safe wrapper for allocateDescriptorSets and freeDescriptorSets using bracket

The allocated value must not be returned from the provided computation

freeDescriptorSets :: Device -> DescriptorPool -> ("descriptorSets" ::: Vector DescriptorSet) -> IO ()

vkFreeDescriptorSets - Free one or more descriptor sets

Parameters

Device is the logical device that owns the descriptor pool.

DescriptorPool is the descriptor pool from which the descriptor sets were allocated.

descriptorSetCount is the number of elements in the pDescriptorSets array.

pDescriptorSets is a pointer to an array of handles to DescriptorSet objects.

Description

After calling freeDescriptorSets, all descriptor sets in pDescriptorSets are invalid.

Valid Usage

All submitted commands that refer to any element of pDescriptorSets must have completed execution

pDescriptorSets must be a valid pointer to an array of descriptorSetCount DescriptorSet handles, each element of which must either be a valid handle or NULL_HANDLE

Each valid handle in pDescriptorSets must have been allocated from DescriptorPool

DescriptorPool must have been created with the DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT flag

Valid Usage (Implicit)

Device must be a valid Device handle

DescriptorPool must be a valid DescriptorPool handle

descriptorSetCount must be greater than 0

DescriptorPool must have been created, allocated, or retrieved from Device

Each element of pDescriptorSets that is a valid handle must have been created, allocated, or retrieved from DescriptorPool

Host Synchronization

Host access to DescriptorPool must be externally synchronized

Host access to each member of pDescriptorSets must be externally synchronized

Return Codes

Success

SUCCESS

See Also

DescriptorPool, DescriptorSet, Device

updateDescriptorSets :: PokeChain a => Device -> ("descriptorWrites" ::: Vector (WriteDescriptorSet a)) -> ("descriptorCopies" ::: Vector CopyDescriptorSet) -> IO ()

vkUpdateDescriptorSets - Update the contents of a descriptor set object

Parameters

Device is the logical device that updates the descriptor sets.

descriptorWriteCount is the number of elements in the pDescriptorWrites array.

pDescriptorWrites is a pointer to an array of WriteDescriptorSet structures describing the descriptor sets to write to.

descriptorCopyCount is the number of elements in the pDescriptorCopies array.

pDescriptorCopies is a pointer to an array of CopyDescriptorSet structures describing the descriptor sets to copy between.

Description

The operations described by pDescriptorWrites are performed first, followed by the operations described by pDescriptorCopies. Within each array, the operations are performed in the order they appear in the array.

Each element in the pDescriptorWrites array describes an operation updating the descriptor set using descriptors for resources specified in the structure.

Each element in the pDescriptorCopies array is a CopyDescriptorSet structure describing an operation copying descriptors between sets.

If the dstSet member of any element of pDescriptorWrites or pDescriptorCopies is bound, accessed, or modified by any command that was recorded to a command buffer which is currently in the recording or executable state, and any of the descriptor bindings that are updated were not created with the DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT or DESCRIPTOR_BINDING_UPDATE_UNUSED_WHILE_PENDING_BIT bits set, that command buffer becomes invalid.

Valid Usage

Descriptor bindings updated by this command which were created without the DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT or DESCRIPTOR_BINDING_UPDATE_UNUSED_WHILE_PENDING_BIT bits set must not be used by any command that was recorded to a command buffer which is in the pending state.

Valid Usage (Implicit)

Device must be a valid Device handle

If descriptorWriteCount is not 0, pDescriptorWrites must be a valid pointer to an array of descriptorWriteCount valid WriteDescriptorSet structures

If descriptorCopyCount is not 0, pDescriptorCopies must be a valid pointer to an array of descriptorCopyCount valid CopyDescriptorSet structures

Host Synchronization

Host access to pDescriptorWrites[].dstSet must be externally synchronized

Host access to pDescriptorCopies[].dstSet must be externally synchronized

See Also

CopyDescriptorSet, Device, WriteDescriptorSet

data DescriptorBufferInfo

VkDescriptorBufferInfo - Structure specifying descriptor buffer info

Description

Note

When setting range to WHOLE_SIZE, the effective range must not be larger than the maximum range for the descriptor type (maxUniformBufferRange or maxStorageBufferRange). This means that WHOLE_SIZE is not typically useful in the common case where uniform buffer descriptors are suballocated from a buffer that is much larger than maxUniformBufferRange.

For DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC and DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC descriptor types, offset is the base offset from which the dynamic offset is applied and range is the static size used for all dynamic offsets.

Valid Usage

offset must be less than the size of Buffer

If range is not equal to WHOLE_SIZE, range must be greater than 0

If range is not equal to WHOLE_SIZE, range must be less than or equal to the size of Buffer minus offset

Valid Usage (Implicit)

Buffer must be a valid Buffer handle

See Also

Buffer, DeviceSize, WriteDescriptorSet

Constructors

DescriptorBufferInfo

Fields

buffer :: Buffer

Buffer is the buffer resource.

offset :: DeviceSize

offset is the offset in bytes from the start of Buffer. Access to buffer memory via this descriptor uses addressing that is relative to this starting offset.

range :: DeviceSize

range is the size in bytes that is used for this descriptor update, or WHOLE_SIZE to use the range from offset to the end of the buffer.

data DescriptorImageInfo

VkDescriptorImageInfo - Structure specifying descriptor image info

Description

Members of DescriptorImageInfo that are not used in an update (as described above) are ignored.

Valid Usage

ImageView must not be 2D or 2D array image view created from a 3D image

If ImageView is created from a depth/stencil image, the aspectMask used to create the ImageView must include either IMAGE_ASPECT_DEPTH_BIT or IMAGE_ASPECT_STENCIL_BIT but not both.

ImageLayout must match the actual ImageLayout of each subresource accessible from ImageView at the time this descriptor is accessed as defined by the image layout matching rules

If Sampler is used and the Format of the image is a multi-planar format, the image must have been created with IMAGE_CREATE_MUTABLE_FORMAT_BIT, and the aspectMask of the ImageView must be IMAGE_ASPECT_PLANE_0_BIT, IMAGE_ASPECT_PLANE_1_BIT or (for three-plane formats only) IMAGE_ASPECT_PLANE_2_BIT

Valid Usage (Implicit)

Both of ImageView, and Sampler that are valid handles of non-ignored parameters must have been created, allocated, or retrieved from the same Device

See Also

ImageLayout, ImageView, Sampler, WriteDescriptorSet

Constructors

DescriptorImageInfo

Fields

sampler :: Sampler

Sampler is a sampler handle, and is used in descriptor updates for types DESCRIPTOR_TYPE_SAMPLER and DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER if the binding being updated does not use immutable samplers.

imageView :: ImageView

ImageView is an image view handle, and is used in descriptor updates for types DESCRIPTOR_TYPE_SAMPLED_IMAGE, DESCRIPTOR_TYPE_STORAGE_IMAGE, DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, and DESCRIPTOR_TYPE_INPUT_ATTACHMENT.

imageLayout :: ImageLayout

ImageLayout is the layout that the image subresources accessible from ImageView will be in at the time this descriptor is accessed. ImageLayout is used in descriptor updates for types DESCRIPTOR_TYPE_SAMPLED_IMAGE, DESCRIPTOR_TYPE_STORAGE_IMAGE, DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, and DESCRIPTOR_TYPE_INPUT_ATTACHMENT.

data WriteDescriptorSet (es :: [Type])

VkWriteDescriptorSet - Structure specifying the parameters of a descriptor set write operation

Description

Only one of pImageInfo, pBufferInfo, or pTexelBufferView members is used according to the descriptor type specified in the DescriptorType member of the containing WriteDescriptorSet structure, or none of them in case DescriptorType is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT, in which case the source data for the descriptor writes is taken from the WriteDescriptorSetInlineUniformBlockEXT structure included in the pNext chain of WriteDescriptorSet, or if DescriptorType is DESCRIPTOR_TYPE_ACCELERATION_STRUCTURE_NV, in which case the source data for the descriptor writes is taken from the WriteDescriptorSetAccelerationStructureNV structure in the pNext chain of WriteDescriptorSet, as specified below.

If the dstBinding has fewer than descriptorCount array elements remaining starting from dstArrayElement, then the remainder will be used to update the subsequent binding - dstBinding+1 starting at array element zero. If a binding has a descriptorCount of zero, it is skipped. This behavior applies recursively, with the update affecting consecutive bindings as needed to update all descriptorCount descriptors.

Note

The same behavior applies to bindings with a descriptor type of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT where descriptorCount specifies the number of bytes to update while dstArrayElement specifies the starting byte offset, thus in this case if the dstBinding has a smaller byte size than the sum of dstArrayElement and descriptorCount, then the remainder will be used to update the subsequent binding - dstBinding+1 starting at offset zero. This falls out as a special case of the above rule.

Valid Usage

dstBinding must be less than or equal to the maximum value of binding of all DescriptorSetLayoutBinding structures specified when dstSet’s descriptor set layout was created

dstBinding must be a binding with a non-zero descriptorCount

All consecutive bindings updated via a single WriteDescriptorSet structure, except those with a descriptorCount of zero, must have identical DescriptorType and stageFlags.

All consecutive bindings updated via a single WriteDescriptorSet structure, except those with a descriptorCount of zero, must all either use immutable samplers or must all not use immutable samplers.

DescriptorType must match the type of dstBinding within dstSet

dstSet must be a valid DescriptorSet handle

The sum of dstArrayElement and descriptorCount must be less than or equal to the number of array elements in the descriptor set binding specified by dstBinding, and all applicable consecutive bindings, as described by https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#descriptorsets-updates-consecutive

If DescriptorType is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT, dstArrayElement must be an integer multiple of 4

If DescriptorType is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT, descriptorCount must be an integer multiple of 4

If DescriptorType is DESCRIPTOR_TYPE_SAMPLER, DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, DESCRIPTOR_TYPE_SAMPLED_IMAGE, DESCRIPTOR_TYPE_STORAGE_IMAGE, or DESCRIPTOR_TYPE_INPUT_ATTACHMENT, pImageInfo must be a valid pointer to an array of descriptorCount valid DescriptorImageInfo structures

If DescriptorType is DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER or DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER, pTexelBufferView must be a valid pointer to an array of descriptorCount valid BufferView handles

If DescriptorType is DESCRIPTOR_TYPE_UNIFORM_BUFFER, DESCRIPTOR_TYPE_STORAGE_BUFFER, DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC, pBufferInfo must be a valid pointer to an array of descriptorCount valid DescriptorBufferInfo structures

If DescriptorType is DESCRIPTOR_TYPE_SAMPLER or DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, and dstSet was not allocated with a layout that included immutable samplers for dstBinding with DescriptorType, the Sampler member of each element of pImageInfo must be a valid Sampler object

If DescriptorType is DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, DESCRIPTOR_TYPE_SAMPLED_IMAGE, DESCRIPTOR_TYPE_STORAGE_IMAGE, or DESCRIPTOR_TYPE_INPUT_ATTACHMENT, the ImageView and ImageLayout members of each element of pImageInfo must be a valid ImageView and ImageLayout, respectively

If DescriptorType is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT, the pNext chain must include a WriteDescriptorSetInlineUniformBlockEXT structure whose dataSize member equals descriptorCount

If DescriptorType is DESCRIPTOR_TYPE_ACCELERATION_STRUCTURE_NV, the pNext chain must include a WriteDescriptorSetAccelerationStructureNV structure whose accelerationStructureCount member equals descriptorCount

If DescriptorType is DESCRIPTOR_TYPE_SAMPLED_IMAGE, then the ImageView member of each pImageInfo element must have been created without a SamplerYcbcrConversionInfo structure in its pNext chain

If DescriptorType is DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, and if any element of pImageInfo has a ImageView member that was created with a SamplerYcbcrConversionInfo structure in its pNext chain, then dstSet must have been allocated with a layout that included immutable samplers for dstBinding, and the corresponding immutable sampler must have been created with an identically defined SamplerYcbcrConversionInfo object

If DescriptorType is DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, and dstSet was allocated with a layout that included immutable samplers for dstBinding, then the ImageView member of each element of pImageInfo which corresponds to an immutable sampler that enables sampler Y′CBCR conversion must have been created with a SamplerYcbcrConversionInfo structure in its pNext chain with an identically defined SamplerYcbcrConversionInfo to the corresponding immutable sampler

If DescriptorType is DESCRIPTOR_TYPE_STORAGE_IMAGE, for each descriptor that will be accessed via load or store operations the ImageLayout member for corresponding elements of pImageInfo must be IMAGE_LAYOUT_GENERAL

If DescriptorType is DESCRIPTOR_TYPE_UNIFORM_BUFFER or DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, the offset member of each element of pBufferInfo must be a multiple of PhysicalDeviceLimits::minUniformBufferOffsetAlignment

If DescriptorType is DESCRIPTOR_TYPE_STORAGE_BUFFER or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC, the offset member of each element of pBufferInfo must be a multiple of PhysicalDeviceLimits::minStorageBufferOffsetAlignment

If DescriptorType is DESCRIPTOR_TYPE_UNIFORM_BUFFER, DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, DESCRIPTOR_TYPE_STORAGE_BUFFER, or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC, and the Buffer member of any element of pBufferInfo is the handle of a non-sparse buffer, then that buffer must be bound completely and contiguously to a single DeviceMemory object

If DescriptorType is DESCRIPTOR_TYPE_UNIFORM_BUFFER or DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, the Buffer member of each element of pBufferInfo must have been created with BUFFER_USAGE_UNIFORM_BUFFER_BIT set

If DescriptorType is DESCRIPTOR_TYPE_STORAGE_BUFFER or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC, the Buffer member of each element of pBufferInfo must have been created with BUFFER_USAGE_STORAGE_BUFFER_BIT set

If DescriptorType is DESCRIPTOR_TYPE_UNIFORM_BUFFER or DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, the range member of each element of pBufferInfo, or the effective range if range is WHOLE_SIZE, must be less than or equal to PhysicalDeviceLimits::maxUniformBufferRange

If DescriptorType is DESCRIPTOR_TYPE_STORAGE_BUFFER or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC, the range member of each element of pBufferInfo, or the effective range if range is WHOLE_SIZE, must be less than or equal to PhysicalDeviceLimits::maxStorageBufferRange

If DescriptorType is DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER, the Buffer that each element of pTexelBufferView was created from must have been created with BUFFER_USAGE_UNIFORM_TEXEL_BUFFER_BIT set

If DescriptorType is DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER, the Buffer that each element of pTexelBufferView was created from must have been created with BUFFER_USAGE_STORAGE_TEXEL_BUFFER_BIT set

If DescriptorType is DESCRIPTOR_TYPE_STORAGE_IMAGE or DESCRIPTOR_TYPE_INPUT_ATTACHMENT, the ImageView member of each element of pImageInfo must have been created with the identity swizzle

If DescriptorType is DESCRIPTOR_TYPE_SAMPLED_IMAGE or DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, the ImageView member of each element of pImageInfo must have been created with IMAGE_USAGE_SAMPLED_BIT set

If DescriptorType is DESCRIPTOR_TYPE_SAMPLED_IMAGE or DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, the ImageLayout member of each element of pImageInfo must be a member of the list given in Sampled Image or Combined Image Sampler, corresponding to its type

If DescriptorType is DESCRIPTOR_TYPE_INPUT_ATTACHMENT, the ImageView member of each element of pImageInfo must have been created with IMAGE_USAGE_INPUT_ATTACHMENT_BIT set

If DescriptorType is DESCRIPTOR_TYPE_STORAGE_IMAGE, the ImageView member of each element of pImageInfo must have been created with IMAGE_USAGE_STORAGE_BIT set

All consecutive bindings updated via a single WriteDescriptorSet structure, except those with a descriptorCount of zero, must have identical DescriptorBindingFlagBits.

If DescriptorType is DESCRIPTOR_TYPE_SAMPLER, then dstSet must not have been allocated with a layout that included immutable samplers for dstBinding

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of WriteDescriptorSetAccelerationStructureNV or WriteDescriptorSetInlineUniformBlockEXT

The sType value of each struct in the pNext chain must be unique

DescriptorType must be a valid DescriptorType value

descriptorCount must be greater than 0

Both of dstSet, and the elements of pTexelBufferView that are valid handles of non-ignored parameters must have been created, allocated, or retrieved from the same Device

See Also

BufferView, DescriptorBufferInfo, DescriptorImageInfo, DescriptorSet, DescriptorType, StructureType, cmdPushDescriptorSetKHR, updateDescriptorSets

Constructors

WriteDescriptorSet

Fields

next :: Chain es

pNext is NULL or a pointer to an extension-specific structure.

dstSet :: DescriptorSet

dstSet is the destination descriptor set to update.

dstBinding :: Word32

dstBinding is the descriptor binding within that set.

dstArrayElement :: Word32

dstArrayElement is the starting element in that array. If the descriptor binding identified by dstSet and dstBinding has a descriptor type of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then dstArrayElement specifies the starting byte offset within the binding.

descriptorType :: DescriptorType

DescriptorType is a DescriptorType specifying the type of each descriptor in pImageInfo, pBufferInfo, or pTexelBufferView, as described below. It must be the same type as that specified in DescriptorSetLayoutBinding for dstSet at dstBinding. The type of the descriptor also controls which array the descriptors are taken from.

imageInfo :: Vector DescriptorImageInfo

pImageInfo is a pointer to an array of DescriptorImageInfo structures or is ignored, as described below.

bufferInfo :: Vector DescriptorBufferInfo

pBufferInfo is a pointer to an array of DescriptorBufferInfo structures or is ignored, as described below.

texelBufferView :: Vector BufferView

pTexelBufferView is a pointer to an array of BufferView handles as described in the Buffer Views section or is ignored, as described below.

data CopyDescriptorSet

VkCopyDescriptorSet - Structure specifying a copy descriptor set operation

Valid Usage

srcBinding must be a valid binding within srcSet

The sum of srcArrayElement and descriptorCount must be less than or equal to the number of array elements in the descriptor set binding specified by srcBinding, and all applicable consecutive bindings, as described by https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#descriptorsets-updates-consecutive

dstBinding must be a valid binding within dstSet

The sum of dstArrayElement and descriptorCount must be less than or equal to the number of array elements in the descriptor set binding specified by dstBinding, and all applicable consecutive bindings, as described by https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#descriptorsets-updates-consecutive

The type of dstBinding within dstSet must be equal to the type of srcBinding within srcSet

If srcSet is equal to dstSet, then the source and destination ranges of descriptors must not overlap, where the ranges may include array elements from consecutive bindings as described by https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#descriptorsets-updates-consecutive

If the descriptor type of the descriptor set binding specified by srcBinding is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT, srcArrayElement must be an integer multiple of 4

If the descriptor type of the descriptor set binding specified by dstBinding is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT, dstArrayElement must be an integer multiple of 4

If the descriptor type of the descriptor set binding specified by either srcBinding or dstBinding is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT, descriptorCount must be an integer multiple of 4

If srcSet’s layout was created with the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT flag set, then dstSet’s layout must also have been created with the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT flag set

If srcSet’s layout was created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT flag set, then dstSet’s layout must also have been created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT flag set

If the descriptor pool from which srcSet was allocated was created with the DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT flag set, then the descriptor pool from which dstSet was allocated must also have been created with the DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT flag set

If the descriptor pool from which srcSet was allocated was created without the DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT flag set, then the descriptor pool from which dstSet was allocated must also have been created without the DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT flag set

If the descriptor type of the descriptor set binding specified by dstBinding is DESCRIPTOR_TYPE_SAMPLER, then dstSet must not have been allocated with a layout that included immutable samplers for dstBinding

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_COPY_DESCRIPTOR_SET

pNext must be NULL

srcSet must be a valid DescriptorSet handle

dstSet must be a valid DescriptorSet handle

Both of dstSet, and srcSet must have been created, allocated, or retrieved from the same Device

See Also

DescriptorSet, StructureType, updateDescriptorSets

Constructors

CopyDescriptorSet

Fields

srcSet :: DescriptorSet

srcSet, srcBinding, and srcArrayElement are the source set, binding, and array element, respectively. If the descriptor binding identified by srcSet and srcBinding has a descriptor type of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then srcArrayElement specifies the starting byte offset within the binding to copy from.

srcBinding :: Word32

srcArrayElement :: Word32

dstSet :: DescriptorSet

dstSet, dstBinding, and dstArrayElement are the destination set, binding, and array element, respectively. If the descriptor binding identified by dstSet and dstBinding has a descriptor type of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then dstArrayElement specifies the starting byte offset within the binding to copy to.

dstBinding :: Word32

dstArrayElement :: Word32

descriptorCount :: Word32

descriptorCount is the number of descriptors to copy from the source to destination. If descriptorCount is greater than the number of remaining array elements in the source or destination binding, those affect consecutive bindings in a manner similar to WriteDescriptorSet above. If the descriptor binding identified by srcSet and srcBinding has a descriptor type of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then descriptorCount specifies the number of bytes to copy and the remaining array elements in the source or destination binding refer to the remaining number of bytes in those.

data DescriptorSetLayoutBinding

VkDescriptorSetLayoutBinding - Structure specifying a descriptor set layout binding

Description

pImmutableSamplers affects initialization of samplers. If DescriptorType specifies a DESCRIPTOR_TYPE_SAMPLER or DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER type descriptor, then pImmutableSamplers can be used to initialize a set of immutable samplers. Immutable samplers are permanently bound into the set layout and must not be changed; updating a DESCRIPTOR_TYPE_SAMPLER descriptor with immutable samplers is not allowed and updates to a DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER descriptor with immutable samplers does not modify the samplers (the image views are updated, but the sampler updates are ignored). If pImmutableSamplers is not NULL, then it points to an array of sampler handles that will be copied into the set layout and used for the corresponding binding. Only the sampler handles are copied; the sampler objects must not be destroyed before the final use of the set layout and any descriptor pools and sets created using it. If pImmutableSamplers is NULL, then the sampler slots are dynamic and sampler handles must be bound into descriptor sets using this layout. If DescriptorType is not one of these descriptor types, then pImmutableSamplers is ignored.

The above layout definition allows the descriptor bindings to be specified sparsely such that not all binding numbers between 0 and the maximum binding number need to be specified in the pBindings array. Bindings that are not specified have a descriptorCount and stageFlags of zero, and the value of DescriptorType is undefined. However, all binding numbers between 0 and the maximum binding number in the DescriptorSetLayoutCreateInfo::pBindings array may consume memory in the descriptor set layout even if not all descriptor bindings are used, though it should not consume additional memory from the descriptor pool.

Note

The maximum binding number specified should be as compact as possible to avoid wasted memory.

Valid Usage

If DescriptorType is DESCRIPTOR_TYPE_SAMPLER or DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, and descriptorCount is not 0 and pImmutableSamplers is not NULL, pImmutableSamplers must be a valid pointer to an array of descriptorCount valid Sampler handles

If DescriptorType is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then descriptorCount must be a multiple of 4

If DescriptorType is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then descriptorCount must be less than or equal to PhysicalDeviceInlineUniformBlockPropertiesEXT::maxInlineUniformBlockSize

If descriptorCount is not 0, stageFlags must be a valid combination of ShaderStageFlagBits values

If DescriptorType is DESCRIPTOR_TYPE_INPUT_ATTACHMENT and descriptorCount is not 0, then stageFlags must be 0 or SHADER_STAGE_FRAGMENT_BIT

Valid Usage (Implicit)

DescriptorType must be a valid DescriptorType value

See Also

DescriptorSetLayoutCreateInfo, DescriptorType, Sampler, ShaderStageFlags

Constructors

DescriptorSetLayoutBinding

Fields

binding :: Word32

binding is the binding number of this entry and corresponds to a resource of the same binding number in the shader stages.

descriptorType :: DescriptorType

DescriptorType is a DescriptorType specifying which type of resource descriptors are used for this binding.

stageFlags :: ShaderStageFlags

stageFlags member is a bitmask of ShaderStageFlagBits specifying which pipeline shader stages can access a resource for this binding. SHADER_STAGE_ALL is a shorthand specifying that all defined shader stages, including any additional stages defined by extensions, can access the resource.

If a shader stage is not included in stageFlags, then a resource must not be accessed from that stage via this binding within any pipeline using the set layout. Other than input attachments which are limited to the fragment shader, there are no limitations on what combinations of stages can use a descriptor binding, and in particular a binding can be used by both graphics stages and the compute stage.

immutableSamplers :: Either Word32 (Vector Sampler)

data DescriptorSetLayoutCreateInfo (es :: [Type])

VkDescriptorSetLayoutCreateInfo - Structure specifying parameters of a newly created descriptor set layout

Valid Usage

The DescriptorSetLayoutBinding::binding members of the elements of the pBindings array must each have different values.

If Flags contains DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR, then all elements of pBindings must not have a DescriptorType of DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC

If Flags contains DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR, then all elements of pBindings must not have a DescriptorType of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT

If Flags contains DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR, then the total number of elements of all bindings must be less than or equal to PhysicalDevicePushDescriptorPropertiesKHR::maxPushDescriptors

If any binding has the DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT bit set, Flags must include DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT

If any binding has the DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT bit set, then all bindings must not have DescriptorType of DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC or DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO

pNext must be NULL or a pointer to a valid instance of DescriptorSetLayoutBindingFlagsCreateInfo

The sType value of each struct in the pNext chain must be unique

Flags must be a valid combination of DescriptorSetLayoutCreateFlagBits values

If bindingCount is not 0, pBindings must be a valid pointer to an array of bindingCount valid DescriptorSetLayoutBinding structures

See Also

DescriptorSetLayoutBinding, DescriptorSetLayoutCreateFlags, StructureType, createDescriptorSetLayout, getDescriptorSetLayoutSupport, getDescriptorSetLayoutSupportKHR

Constructors

DescriptorSetLayoutCreateInfo

Fields

next :: Chain es

pNext is NULL or a pointer to an extension-specific structure.

flags :: DescriptorSetLayoutCreateFlags

Flags is a bitmask of DescriptorSetLayoutCreateFlagBits specifying options for descriptor set layout creation.

bindings :: Vector DescriptorSetLayoutBinding

pBindings is a pointer to an array of DescriptorSetLayoutBinding structures.

data DescriptorPoolSize

VkDescriptorPoolSize - Structure specifying descriptor pool size

Description

Note

When creating a descriptor pool that will contain descriptors for combined image samplers of multi-planar formats, an application needs to account for non-trivial descriptor consumption when choosing the descriptorCount value, as indicated by SamplerYcbcrConversionImageFormatProperties::combinedImageSamplerDescriptorCount.

Valid Usage

descriptorCount must be greater than 0

If type is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then descriptorCount must be a multiple of 4

Valid Usage (Implicit)

type must be a valid DescriptorType value

See Also

DescriptorPoolCreateInfo, DescriptorType

Constructors

DescriptorPoolSize

Fields

type' :: DescriptorType

type is the type of descriptor.

descriptorCount :: Word32

descriptorCount is the number of descriptors of that type to allocate. If type is DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT then descriptorCount is the number of bytes to allocate for descriptors of this type.

data DescriptorPoolCreateInfo (es :: [Type])

VkDescriptorPoolCreateInfo - Structure specifying parameters of a newly created descriptor pool

Description

If multiple DescriptorPoolSize structures appear in the pPoolSizes array then the pool will be created with enough storage for the total number of descriptors of each type.

Fragmentation of a descriptor pool is possible and may lead to descriptor set allocation failures. A failure due to fragmentation is defined as failing a descriptor set allocation despite the sum of all outstanding descriptor set allocations from the pool plus the requested allocation requiring no more than the total number of descriptors requested at pool creation. Implementations provide certain guarantees of when fragmentation must not cause allocation failure, as described below.

If a descriptor pool has not had any descriptor sets freed since it was created or most recently reset then fragmentation must not cause an allocation failure (note that this is always the case for a pool created without the DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT bit set). Additionally, if all sets allocated from the pool since it was created or most recently reset use the same number of descriptors (of each type) and the requested allocation also uses that same number of descriptors (of each type), then fragmentation must not cause an allocation failure.

If an allocation failure occurs due to fragmentation, an application can create an additional descriptor pool to perform further descriptor set allocations.

If Flags has the DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT bit set, descriptor pool creation may fail with the error ERROR_FRAGMENTATION if the total number of descriptors across all pools (including this one) created with this bit set exceeds maxUpdateAfterBindDescriptorsInAllPools, or if fragmentation of the underlying hardware resources occurs.

Valid Usage

maxSets must be greater than 0

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO

pNext must be NULL or a pointer to a valid instance of DescriptorPoolInlineUniformBlockCreateInfoEXT

The sType value of each struct in the pNext chain must be unique

Flags must be a valid combination of DescriptorPoolCreateFlagBits values

pPoolSizes must be a valid pointer to an array of poolSizeCount valid DescriptorPoolSize structures

poolSizeCount must be greater than 0

See Also

DescriptorPoolCreateFlags, DescriptorPoolSize, StructureType, createDescriptorPool

Constructors

DescriptorPoolCreateInfo

Fields

next :: Chain es

pNext is NULL or a pointer to an extension-specific structure.

flags :: DescriptorPoolCreateFlags

Flags is a bitmask of DescriptorPoolCreateFlagBits specifying certain supported operations on the pool.

maxSets :: Word32

maxSets is the maximum number of descriptor sets that can be allocated from the pool.

poolSizes :: Vector DescriptorPoolSize

pPoolSizes is a pointer to an array of DescriptorPoolSize structures, each containing a descriptor type and number of descriptors of that type to be allocated in the pool.

data DescriptorSetAllocateInfo (es :: [Type])

VkDescriptorSetAllocateInfo - Structure specifying the allocation parameters for descriptor sets

Valid Usage

Each element of pSetLayouts must not have been created with DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR set

If any element of pSetLayouts was created with the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set, DescriptorPool must have been created with the DESCRIPTOR_POOL_CREATE_UPDATE_AFTER_BIND_BIT flag set

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO

pNext must be NULL or a pointer to a valid instance of DescriptorSetVariableDescriptorCountAllocateInfo

The sType value of each struct in the pNext chain must be unique

DescriptorPool must be a valid DescriptorPool handle

pSetLayouts must be a valid pointer to an array of descriptorSetCount valid DescriptorSetLayout handles

descriptorSetCount must be greater than 0

Both of DescriptorPool, and the elements of pSetLayouts must have been created, allocated, or retrieved from the same Device

See Also

DescriptorPool, DescriptorSetLayout, StructureType, allocateDescriptorSets

Constructors

DescriptorSetAllocateInfo

Fields

next :: Chain es

pNext is NULL or a pointer to an extension-specific structure.

descriptorPool :: DescriptorPool

DescriptorPool is the pool which the sets will be allocated from.

setLayouts :: Vector DescriptorSetLayout

pSetLayouts is a pointer to an array of descriptor set layouts, with each member specifying how the corresponding descriptor set is allocated.