Graphics.Vulkan.Core10.ImageView

Documentation

createImageView :: PokeChain a => Device -> ImageViewCreateInfo a -> ("allocator" ::: Maybe AllocationCallbacks) -> IO ImageView

vkCreateImageView - Create an image view from an existing image

Parameters

Device is the logical device that creates the image view.

pCreateInfo is a pointer to a ImageViewCreateInfo structure containing parameters to be used to create the image view.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pView is a pointer to a ImageView handle in which the resulting image view object is returned.

Valid Usage (Implicit)

Device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid ImageViewCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pView must be a valid pointer to a ImageView handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, Device, ImageView, ImageViewCreateInfo

withImageView :: PokeChain a => Device -> ImageViewCreateInfo a -> Maybe AllocationCallbacks -> (ImageView -> IO r) -> IO r

A safe wrapper for createImageView and destroyImageView using bracket

The allocated value must not be returned from the provided computation

destroyImageView :: Device -> ImageView -> ("allocator" ::: Maybe AllocationCallbacks) -> IO ()

vkDestroyImageView - Destroy an image view object

Parameters

Device is the logical device that destroys the image view.

ImageView is the image view to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to ImageView must have completed execution

If AllocationCallbacks were provided when ImageView was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when ImageView was created, pAllocator must be NULL

Valid Usage (Implicit)

Device must be a valid Device handle

If ImageView is not NULL_HANDLE, ImageView must be a valid ImageView handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If ImageView is a valid handle, it must have been created, allocated, or retrieved from Device

Host Synchronization

Host access to ImageView must be externally synchronized

See Also

AllocationCallbacks, Device, ImageView

data ComponentMapping

VkComponentMapping - Structure specifying a color component mapping

Valid Usage (Implicit)

See Also

AndroidHardwareBufferFormatPropertiesANDROID, ComponentSwizzle, ImageViewCreateInfo, SamplerYcbcrConversionCreateInfo

Constructors

ComponentMapping

Fields

r :: ComponentSwizzle

r must be a valid ComponentSwizzle value

g :: ComponentSwizzle

g must be a valid ComponentSwizzle value

b :: ComponentSwizzle

b must be a valid ComponentSwizzle value

a :: ComponentSwizzle

a must be a valid ComponentSwizzle value

data ImageViewCreateInfo (es :: [Type])

VkImageViewCreateInfo - Structure specifying parameters of a newly created image view

Description

Some of the Image creation parameters are inherited by the view. In particular, image view creation inherits the implicit parameter usage specifying the allowed usages of the image view that, by default, takes the value of the corresponding usage parameter specified in ImageCreateInfo at image creation time. If the image was has a depth-stencil format and was created with a ImageStencilUsageCreateInfo structure included in the pNext chain of ImageCreateInfo, the usage is calculated based on the subresource.aspectMask provided:

If aspectMask includes only IMAGE_ASPECT_STENCIL_BIT, the implicit usage is equal to ImageStencilUsageCreateInfo::stencilUsage.

If aspectMask includes only IMAGE_ASPECT_DEPTH_BIT, the implicit usage is equal to ImageCreateInfo::usage.

If both aspects are included in aspectMask, the implicit usage is equal to the intersection of ImageCreateInfo::usage and ImageStencilUsageCreateInfo::stencilUsage. The implicit usage can be overriden by adding a ImageViewUsageCreateInfo structure to the pNext chain.

If Image was created with the IMAGE_CREATE_MUTABLE_FORMAT_BIT flag, and if the Format of the image is not multi-planar, Format can be different from the image’s format, but if Image was created without the IMAGE_CREATE_BLOCK_TEXEL_VIEW_COMPATIBLE_BIT flag and they are not equal they must be compatible. Image format compatibility is defined in the Format Compatibility Classes section. Views of compatible formats will have the same mapping between texel coordinates and memory locations irrespective of the Format, with only the interpretation of the bit pattern changing.

Note

Values intended to be used with one view format may not be exactly preserved when written or read through a different format. For example, an integer value that happens to have the bit pattern of a floating point denorm or NaN may be flushed or canonicalized when written or read through a view with a floating point format. Similarly, a value written through a signed normalized format that has a bit pattern exactly equal to -2b may be changed to -2b + 1 as described in Conversion from Normalized Fixed-Point to Floating-Point.

If Image was created with the IMAGE_CREATE_BLOCK_TEXEL_VIEW_COMPATIBLE_BIT flag, Format must be compatible with the image’s format as described above, or must be an uncompressed format in which case it must be size-compatible with the image’s format, as defined for copying data between images In this case the resulting image view’s texel dimensions equal the dimensions of the selected mip level divided by the compressed texel block size and rounded up.

If the image view is to be used with a sampler which supports sampler Y′CBCR conversion, an identically defined object of type SamplerYcbcrConversion to that used to create the sampler must be passed to createImageView in a SamplerYcbcrConversionInfo included in the pNext chain of ImageViewCreateInfo. Conversely, if a SamplerYcbcrConversion object is passed to createImageView, an identically defined SamplerYcbcrConversion object must be used when sampling the image.

If the image has a multi-planar Format and subresourceRange.aspectMask is IMAGE_ASPECT_COLOR_BIT, Format must be identical to the image Format, and the sampler to be used with the image view must enable sampler Y′CBCR conversion.

If Image was created with the IMAGE_CREATE_MUTABLE_FORMAT_BIT and the image has a multi-planar Format, and if subresourceRange.aspectMask is IMAGE_ASPECT_PLANE_0_BIT, IMAGE_ASPECT_PLANE_1_BIT, or IMAGE_ASPECT_PLANE_2_BIT, Format must be compatible with the corresponding plane of the image, and the sampler to be used with the image view must not enable sampler Y′CBCR conversion. The width and height of the single-plane image view must be derived from the multi-planar image’s dimensions in the manner listed for plane compatibility for the plane.

Any view of an image plane will have the same mapping between texel coordinates and memory locations as used by the channels of the color aspect, subject to the formulae relating texel coordinates to lower-resolution planes as described in Chroma Reconstruction. That is, if an R or B plane has a reduced resolution relative to the G plane of the multi-planar image, the image view operates using the (uplane, vplane) unnormalized coordinates of the reduced-resolution plane, and these coordinates access the same memory locations as the (ucolor, vcolor) unnormalized coordinates of the color aspect for which chroma reconstruction operations operate on the same (uplane, vplane) or (iplane, jplane) coordinates.

Dim, Arrayed, MS 	 Image parameters 	 View parameters

ImageType = ci.ImageType width = ci.extent.width height = ci.extent.height depth = ci.extent.depth arrayLayers = ci.arrayLayers samples = ci.samples Flags = ci.Flags where ci is the ImageCreateInfo used to create Image. 	 baseArrayLayer, layerCount, and levelCount are members of the subresourceRange member.

__1D, 0, 0__ 	 ImageType = IMAGE_TYPE_1D width ≥ 1 height = 1 depth = 1 arrayLayers ≥ 1 samples = 1 	 viewType = IMAGE_VIEW_TYPE_1D baseArrayLayer ≥ 0 layerCount = 1

__1D, 1, 0__ 	 ImageType = IMAGE_TYPE_1D width ≥ 1 height = 1 depth = 1 arrayLayers ≥ 1 samples = 1 	 viewType = IMAGE_VIEW_TYPE_1D_ARRAY baseArrayLayer ≥ 0 layerCount ≥ 1

__2D, 0, 0__ 	 ImageType = IMAGE_TYPE_2D width ≥ 1 height ≥ 1 depth = 1 arrayLayers ≥ 1 samples = 1 	 viewType = IMAGE_VIEW_TYPE_2D baseArrayLayer ≥ 0 layerCount = 1

__2D, 1, 0__ 	 ImageType = IMAGE_TYPE_2D width ≥ 1 height ≥ 1 depth = 1 arrayLayers ≥ 1 samples = 1 	 viewType = IMAGE_VIEW_TYPE_2D_ARRAY baseArrayLayer ≥ 0 layerCount ≥ 1

__2D, 0, 1__ 	 ImageType = IMAGE_TYPE_2D width ≥ 1 height ≥ 1 depth = 1 arrayLayers ≥ 1 samples > 1 	 viewType = IMAGE_VIEW_TYPE_2D baseArrayLayer ≥ 0 layerCount = 1

__2D, 1, 1__ 	 ImageType = IMAGE_TYPE_2D width ≥ 1 height ≥ 1 depth = 1 arrayLayers ≥ 1 samples > 1 	 viewType = IMAGE_VIEW_TYPE_2D_ARRAY baseArrayLayer ≥ 0 layerCount ≥ 1

__CUBE, 0, 0__ 	 ImageType = IMAGE_TYPE_2D width ≥ 1 height = width depth = 1 arrayLayers ≥ 6 samples = 1 Flags includes IMAGE_CREATE_CUBE_COMPATIBLE_BIT 	 viewType = IMAGE_VIEW_TYPE_CUBE baseArrayLayer ≥ 0 layerCount = 6

__CUBE, 1, 0__ 	 ImageType = IMAGE_TYPE_2D width ≥ 1 height = width depth = 1 N ≥ 1 arrayLayers ≥ 6 × N samples = 1 Flags includes IMAGE_CREATE_CUBE_COMPATIBLE_BIT 	 viewType = IMAGE_VIEW_TYPE_CUBE_ARRAY baseArrayLayer ≥ 0 layerCount = 6 × N, N ≥ 1

__3D, 0, 0__ 	 ImageType = IMAGE_TYPE_3D width ≥ 1 height ≥ 1 depth ≥ 1 arrayLayers = 1 samples = 1 	 viewType = IMAGE_VIEW_TYPE_3D baseArrayLayer = 0 layerCount = 1

__3D, 0, 0__ 	 ImageType = IMAGE_TYPE_3D width ≥ 1 height ≥ 1 depth ≥ 1 arrayLayers = 1 samples = 1 Flags includes IMAGE_CREATE_2D_ARRAY_COMPATIBLE_BIT Flags does not include IMAGE_CREATE_SPARSE_BINDING_BIT, IMAGE_CREATE_SPARSE_RESIDENCY_BIT, and IMAGE_CREATE_SPARSE_ALIASED_BIT 	 viewType = IMAGE_VIEW_TYPE_2D levelCount = 1 baseArrayLayer ≥ 0 layerCount = 1

__3D, 0, 0__ 	 ImageType = IMAGE_TYPE_3D width ≥ 1 height ≥ 1 depth ≥ 1 arrayLayers = 1 samples = 1 Flags includes IMAGE_CREATE_2D_ARRAY_COMPATIBLE_BIT Flags does not include IMAGE_CREATE_SPARSE_BINDING_BIT, IMAGE_CREATE_SPARSE_RESIDENCY_BIT, and IMAGE_CREATE_SPARSE_ALIASED_BIT 	 viewType = IMAGE_VIEW_TYPE_2D_ARRAY levelCount = 1 baseArrayLayer ≥ 0 layerCount ≥ 1

Image and image view parameter compatibility requirements

Valid Usage

If Image was not created with IMAGE_CREATE_CUBE_COMPATIBLE_BIT then viewType must not be IMAGE_VIEW_TYPE_CUBE or IMAGE_VIEW_TYPE_CUBE_ARRAY

If the image cubemap arrays feature is not enabled, viewType must not be IMAGE_VIEW_TYPE_CUBE_ARRAY

If Image was created with IMAGE_TYPE_3D but without IMAGE_CREATE_2D_ARRAY_COMPATIBLE_BIT set then viewType must not be IMAGE_VIEW_TYPE_2D or IMAGE_VIEW_TYPE_2D_ARRAY

Image must have been created with a usage value containing at least one of IMAGE_USAGE_SAMPLED_BIT, IMAGE_USAGE_STORAGE_BIT, IMAGE_USAGE_COLOR_ATTACHMENT_BIT, IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT, IMAGE_USAGE_INPUT_ATTACHMENT_BIT, IMAGE_USAGE_SHADING_RATE_IMAGE_BIT_NV, or IMAGE_USAGE_FRAGMENT_DENSITY_MAP_BIT_EXT

The format features of the resultant image view must contain at least one bit.

If usage contains IMAGE_USAGE_SAMPLED_BIT, then the format features of the resultant image view must contain FORMAT_FEATURE_SAMPLED_IMAGE_BIT.

If usage contains IMAGE_USAGE_STORAGE_BIT, then the image view’s format features must contain FORMAT_FEATURE_STORAGE_IMAGE_BIT.

If usage contains IMAGE_USAGE_COLOR_ATTACHMENT_BIT, then the image view’s format features must contain FORMAT_FEATURE_COLOR_ATTACHMENT_BIT.

If usage contains IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT, then the image view’s format features must contain FORMAT_FEATURE_DEPTH_STENCIL_ATTACHMENT_BIT.

If usage contains IMAGE_USAGE_INPUT_ATTACHMENT_BIT, then the image view’s format features must contain at least one of FORMAT_FEATURE_COLOR_ATTACHMENT_BIT or FORMAT_FEATURE_DEPTH_STENCIL_ATTACHMENT_BIT.

subresourceRange.baseMipLevel must be less than the mipLevels specified in ImageCreateInfo when Image was created

If subresourceRange.levelCount is not REMAINING_MIP_LEVELS, subresourceRange.baseMipLevel + subresourceRange.levelCount must be less than or equal to the mipLevels specified in ImageCreateInfo when Image was created

If Image was created with usage containing IMAGE_USAGE_FRAGMENT_DENSITY_MAP_BIT_EXT, subresourceRange.levelCount must be 1

If Image is not a 3D image created with IMAGE_CREATE_2D_ARRAY_COMPATIBLE_BIT set, or viewType is not IMAGE_VIEW_TYPE_2D or IMAGE_VIEW_TYPE_2D_ARRAY, subresourceRange.baseArrayLayer must be less than the arrayLayers specified in ImageCreateInfo when Image was created

If subresourceRange.layerCount is not REMAINING_ARRAY_LAYERS, Image is not a 3D image created with IMAGE_CREATE_2D_ARRAY_COMPATIBLE_BIT set, or viewType is not IMAGE_VIEW_TYPE_2D or IMAGE_VIEW_TYPE_2D_ARRAY, subresourceRange.layerCount must be non-zero and subresourceRange.baseArrayLayer + subresourceRange.layerCount must be less than or equal to the arrayLayers specified in ImageCreateInfo when Image was created

If Image is a 3D image created with IMAGE_CREATE_2D_ARRAY_COMPATIBLE_BIT set, and viewType is IMAGE_VIEW_TYPE_2D or IMAGE_VIEW_TYPE_2D_ARRAY, subresourceRange.baseArrayLayer must be less than the depth computed from baseMipLevel and extent.depth specified in ImageCreateInfo when Image was created, according to the formula defined in Image Miplevel Sizing.

If subresourceRange.layerCount is not REMAINING_ARRAY_LAYERS, Image is a 3D image created with IMAGE_CREATE_2D_ARRAY_COMPATIBLE_BIT set, and viewType is IMAGE_VIEW_TYPE_2D or IMAGE_VIEW_TYPE_2D_ARRAY, subresourceRange.layerCount must be non-zero and subresourceRange.baseArrayLayer + subresourceRange.layerCount must be less than or equal to the depth computed from baseMipLevel and extent.depth specified in ImageCreateInfo when Image was created, according to the formula defined in Image Miplevel Sizing.

If Image was created with the IMAGE_CREATE_MUTABLE_FORMAT_BIT flag, Format must be compatible with the Format used to create Image, as defined in Format Compatibility Classes

If Image was created with the IMAGE_CREATE_MUTABLE_FORMAT_BIT flag, but without the IMAGE_CREATE_BLOCK_TEXEL_VIEW_COMPATIBLE_BIT flag, and if the Format of the Image is not a multi-planar format, Format must be compatible with the Format used to create Image, as defined in Format Compatibility Classes

If Image was created with the IMAGE_CREATE_BLOCK_TEXEL_VIEW_COMPATIBLE_BIT flag, Format must be compatible with, or must be an uncompressed format that is size-compatible with, the Format used to create Image.

If Image was created with the IMAGE_CREATE_BLOCK_TEXEL_VIEW_COMPATIBLE_BIT flag, the levelCount and layerCount members of subresourceRange must both be 1.

If a ImageFormatListCreateInfo structure was included in the pNext chain of the ImageCreateInfo structure used when creating Image and the viewFormatCount field of ImageFormatListCreateInfo is not zero then Format must be one of the formats in ImageFormatListCreateInfo::pViewFormats.

If Image was created with the IMAGE_CREATE_MUTABLE_FORMAT_BIT flag, if the Format of the Image is a multi-planar format, and if subresourceRange.aspectMask is one of IMAGE_ASPECT_PLANE_0_BIT, IMAGE_ASPECT_PLANE_1_BIT, or IMAGE_ASPECT_PLANE_2_BIT, then Format must be compatible with the Format for the plane of the Image Format indicated by subresourceRange.aspectMask, as defined in https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#formats-compatible-planes

If Image was not created with the IMAGE_CREATE_MUTABLE_FORMAT_BIT flag, or if the Format of the Image is a multi-planar format and if subresourceRange.aspectMask is IMAGE_ASPECT_COLOR_BIT, Format must be identical to the Format used to create Image

If the pNext chain includes a SamplerYcbcrConversionInfo structure with a conversion value other than NULL_HANDLE, all members of components must have the value COMPONENT_SWIZZLE_IDENTITY.

If Image is non-sparse then it must be bound completely and contiguously to a single DeviceMemory object

subresourceRange and viewType must be compatible with the image, as described in the compatibility table

If Image has an external format, Format must be FORMAT_UNDEFINED.

If Image has an external format, the pNext chain must include a SamplerYcbcrConversionInfo structure with a conversion object created with the same external format as Image.

If Image has an external format, all members of components must be COMPONENT_SWIZZLE_IDENTITY.

If Image was created with usage containing IMAGE_USAGE_SHADING_RATE_IMAGE_BIT_NV, viewType must be IMAGE_VIEW_TYPE_2D or IMAGE_VIEW_TYPE_2D_ARRAY

If Image was created with usage containing IMAGE_USAGE_SHADING_RATE_IMAGE_BIT_NV, Format must be FORMAT_R8_UINT

If dynamic fragment density map feature is not enabled, Flags must not contain IMAGE_VIEW_CREATE_FRAGMENT_DENSITY_MAP_DYNAMIC_BIT_EXT

If dynamic fragment density map feature is not enabled and Image was created with usage containing IMAGE_USAGE_FRAGMENT_DENSITY_MAP_BIT_EXT, Flags must not contain any of IMAGE_CREATE_PROTECTED_BIT, IMAGE_CREATE_SPARSE_BINDING_BIT, IMAGE_CREATE_SPARSE_RESIDENCY_BIT, or IMAGE_CREATE_SPARSE_ALIASED_BIT

If the pNext chain includes a ImageViewUsageCreateInfo structure, and Image was not created with a ImageStencilUsageCreateInfo structure included in the pNext chain of ImageCreateInfo, its usage member must not include any bits that were not set in the usage member of the ImageCreateInfo structure used to create Image

If the pNext chain includes a ImageViewUsageCreateInfo structure, Image was created with a ImageStencilUsageCreateInfo structure included in the pNext chain of ImageCreateInfo, and subResourceRange.aspectMask includes IMAGE_ASPECT_STENCIL_BIT, the usage member of the ImageViewUsageCreateInfo instance must not include any bits that were not set in the usage member of the ImageStencilUsageCreateInfo structure used to create Image

If the pNext chain includes a ImageViewUsageCreateInfo structure, Image was created with a ImageStencilUsageCreateInfo structure included in the pNext chain of ImageCreateInfo, and subResourceRange.aspectMask includes bits other than IMAGE_ASPECT_STENCIL_BIT, the usage member of the ImageViewUsageCreateInfo structure must not include any bits that were not set in the usage member of the ImageCreateInfo structure used to create Image

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of ImageViewASTCDecodeModeEXT, ImageViewUsageCreateInfo, or SamplerYcbcrConversionInfo

The sType value of each struct in the pNext chain must be unique

Flags must be a valid combination of ImageViewCreateFlagBits values

Image must be a valid Image handle

viewType must be a valid ImageViewType value

Format must be a valid Format value

components must be a valid ComponentMapping structure

subresourceRange must be a valid ImageSubresourceRange structure

See Also

ComponentMapping, Format, Image, ImageSubresourceRange, ImageViewCreateFlags, ImageViewType, StructureType, createImageView

Constructors

ImageViewCreateInfo

Fields

next :: Chain es

pNext is NULL or a pointer to an extension-specific structure.

flags :: ImageViewCreateFlags

Flags is a bitmask of ImageViewCreateFlagBits describing additional parameters of the image view.

image :: Image

Image is a Image on which the view will be created.

viewType :: ImageViewType

viewType is a ImageViewType value specifying the type of the image view.

format :: Format

Format is a Format describing the format and type used to interpret texel blocks in the image.

components :: ComponentMapping

components is a ComponentMapping specifies a remapping of color components (or of depth or stencil components after they have been converted into color components).

subresourceRange :: ImageSubresourceRange

subresourceRange is a ImageSubresourceRange selecting the set of mipmap levels and array layers to be accessible to the view.