Graphics.Vulkan.Core10.Pass

Documentation

createFramebuffer :: forall a io. (PokeChain a, MonadIO io) => Device -> FramebufferCreateInfo a -> ("allocator" ::: Maybe AllocationCallbacks) -> io Framebuffer

vkCreateFramebuffer - Create a new framebuffer object

Parameters

device is the logical device that creates the framebuffer.

pCreateInfo is a pointer to a FramebufferCreateInfo structure describing additional information about framebuffer creation.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pFramebuffer is a pointer to a Framebuffer handle in which the resulting framebuffer object is returned.

Valid Usage

If pCreateInfo->flags does not include FRAMEBUFFER_CREATE_IMAGELESS_BIT, and attachmentCount is not 0, each element of pCreateInfo->pAttachments must have been created on device

Valid Usage (Implicit)

device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid FramebufferCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pFramebuffer must be a valid pointer to a Framebuffer handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, Device, Framebuffer, FramebufferCreateInfo

withFramebuffer :: forall a r. PokeChain a => Device -> FramebufferCreateInfo a -> Maybe AllocationCallbacks -> (Framebuffer -> IO r) -> IO r

A safe wrapper for createFramebuffer and destroyFramebuffer using bracket

The allocated value must not be returned from the provided computation

destroyFramebuffer :: forall io. MonadIO io => Device -> Framebuffer -> ("allocator" ::: Maybe AllocationCallbacks) -> io ()

vkDestroyFramebuffer - Destroy a framebuffer object

Parameters

device is the logical device that destroys the framebuffer.

framebuffer is the handle of the framebuffer to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to framebuffer must have completed execution

If AllocationCallbacks were provided when framebuffer was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when framebuffer was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid Device handle

If framebuffer is not NULL_HANDLE, framebuffer must be a valid Framebuffer handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If framebuffer is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to framebuffer must be externally synchronized

See Also

AllocationCallbacks, Device, Framebuffer

createRenderPass :: forall a io. (PokeChain a, MonadIO io) => Device -> RenderPassCreateInfo a -> ("allocator" ::: Maybe AllocationCallbacks) -> io RenderPass

vkCreateRenderPass - Create a new render pass object

Parameters

device is the logical device that creates the render pass.

pCreateInfo is a pointer to a RenderPassCreateInfo structure describing the parameters of the render pass.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pRenderPass is a pointer to a RenderPass handle in which the resulting render pass object is returned.

Valid Usage (Implicit)

device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid RenderPassCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pRenderPass must be a valid pointer to a RenderPass handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, Device, RenderPass, RenderPassCreateInfo

withRenderPass :: forall a r. PokeChain a => Device -> RenderPassCreateInfo a -> Maybe AllocationCallbacks -> (RenderPass -> IO r) -> IO r

A safe wrapper for createRenderPass and destroyRenderPass using bracket

The allocated value must not be returned from the provided computation

destroyRenderPass :: forall io. MonadIO io => Device -> RenderPass -> ("allocator" ::: Maybe AllocationCallbacks) -> io ()

vkDestroyRenderPass - Destroy a render pass object

Parameters

device is the logical device that destroys the render pass.

renderPass is the handle of the render pass to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to renderPass must have completed execution

If AllocationCallbacks were provided when renderPass was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when renderPass was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid Device handle

If renderPass is not NULL_HANDLE, renderPass must be a valid RenderPass handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If renderPass is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to renderPass must be externally synchronized

See Also

AllocationCallbacks, Device, RenderPass

getRenderAreaGranularity :: forall io. MonadIO io => Device -> RenderPass -> io ("granularity" ::: Extent2D)

vkGetRenderAreaGranularity - Returns the granularity for optimal render area

Parameters

device is the logical device that owns the render pass.

renderPass is a handle to a render pass.

pGranularity is a pointer to a Extent2D structure in which the granularity is returned.

Description

The conditions leading to an optimal renderArea are:

the offset.x member in renderArea is a multiple of the width member of the returned Extent2D (the horizontal granularity).

the offset.y member in renderArea is a multiple of the height of the returned Extent2D (the vertical granularity).

either the offset.width member in renderArea is a multiple of the horizontal granularity or offset.x+offset.width is equal to the width of the framebuffer in the RenderPassBeginInfo.

either the offset.height member in renderArea is a multiple of the vertical granularity or offset.y+offset.height is equal to the height of the framebuffer in the RenderPassBeginInfo.

Subpass dependencies are not affected by the render area, and apply to the entire image subresources attached to the framebuffer as specified in the description of automatic layout transitions. Similarly, pipeline barriers are valid even if their effect extends outside the render area.

Valid Usage (Implicit)

device must be a valid Device handle

renderPass must be a valid RenderPass handle

pGranularity must be a valid pointer to a Extent2D structure

renderPass must have been created, allocated, or retrieved from device

See Also

Device, Extent2D, RenderPass

data AttachmentDescription

VkAttachmentDescription - Structure specifying an attachment description

Description

If the attachment uses a color format, then loadOp and storeOp are used, and stencilLoadOp and stencilStoreOp are ignored. If the format has depth and/or stencil components, loadOp and storeOp apply only to the depth data, while stencilLoadOp and stencilStoreOp define how the stencil data is handled. loadOp and stencilLoadOp define the load operations that execute as part of the first subpass that uses the attachment. storeOp and stencilStoreOp define the store operations that execute as part of the last subpass that uses the attachment.

The load operation for each sample in an attachment happens-before any recorded command which accesses the sample in the first subpass where the attachment is used. Load operations for attachments with a depth/stencil format execute in the PIPELINE_STAGE_EARLY_FRAGMENT_TESTS_BIT pipeline stage. Load operations for attachments with a color format execute in the PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT pipeline stage.

The store operation for each sample in an attachment happens-after any recorded command which accesses the sample in the last subpass where the attachment is used. Store operations for attachments with a depth/stencil format execute in the PIPELINE_STAGE_LATE_FRAGMENT_TESTS_BIT pipeline stage. Store operations for attachments with a color format execute in the PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT pipeline stage.

If an attachment is not used by any subpass, then loadOp, storeOp, stencilStoreOp, and stencilLoadOp are ignored, and the attachment’s memory contents will not be modified by execution of a render pass instance.

The load and store operations apply on the first and last use of each view in the render pass, respectively. If a view index of an attachment is not included in the view mask in any subpass that uses it, then the load and store operations are ignored, and the attachment’s memory contents will not be modified by execution of a render pass instance.

During a render pass instance, input/color attachments with color formats that have a component size of 8, 16, or 32 bits must be represented in the attachment’s format throughout the instance. Attachments with other floating- or fixed-point color formats, or with depth components may be represented in a format with a precision higher than the attachment format, but must be represented with the same range. When such a component is loaded via the loadOp, it will be converted into an implementation-dependent format used by the render pass. Such components must be converted from the render pass format, to the format of the attachment, before they are resolved or stored at the end of a render pass instance via storeOp. Conversions occur as described in Numeric Representation and Computation and Fixed-Point Data Conversions.

If flags includes ATTACHMENT_DESCRIPTION_MAY_ALIAS_BIT, then the attachment is treated as if it shares physical memory with another attachment in the same render pass. This information limits the ability of the implementation to reorder certain operations (like layout transitions and the loadOp) such that it is not improperly reordered against other uses of the same physical memory via a different attachment. This is described in more detail below.

If a render pass uses multiple attachments that alias the same device memory, those attachments must each include the ATTACHMENT_DESCRIPTION_MAY_ALIAS_BIT bit in their attachment description flags. Attachments aliasing the same memory occurs in multiple ways:

Multiple attachments being assigned the same image view as part of framebuffer creation.

Attachments using distinct image views that correspond to the same image subresource of an image.

Attachments using views of distinct image subresources which are bound to overlapping memory ranges.

Note

Render passes must include subpass dependencies (either directly or via a subpass dependency chain) between any two subpasses that operate on the same attachment or aliasing attachments and those subpass dependencies must include execution and memory dependencies separating uses of the aliases, if at least one of those subpasses writes to one of the aliases. These dependencies must not include the DEPENDENCY_BY_REGION_BIT if the aliases are views of distinct image subresources which overlap in memory.

Multiple attachments that alias the same memory must not be used in a single subpass. A given attachment index must not be used multiple times in a single subpass, with one exception: two subpass attachments can use the same attachment index if at least one use is as an input attachment and neither use is as a resolve or preserve attachment. In other words, the same view can be used simultaneously as an input and color or depth/stencil attachment, but must not be used as multiple color or depth/stencil attachments nor as resolve or preserve attachments. The precise set of valid scenarios is described in more detail below.

If a set of attachments alias each other, then all except the first to be used in the render pass must use an initialLayout of IMAGE_LAYOUT_UNDEFINED, since the earlier uses of the other aliases make their contents undefined. Once an alias has been used and a different alias has been used after it, the first alias must not be used in any later subpasses. However, an application can assign the same image view to multiple aliasing attachment indices, which allows that image view to be used multiple times even if other aliases are used in between.

Note

Once an attachment needs the ATTACHMENT_DESCRIPTION_MAY_ALIAS_BIT bit, there should be no additional cost of introducing additional aliases, and using these additional aliases may allow more efficient clearing of the attachments on multiple uses via ATTACHMENT_LOAD_OP_CLEAR.

Valid Usage

finalLayout must not be IMAGE_LAYOUT_UNDEFINED or IMAGE_LAYOUT_PREINITIALIZED

If format is a color format, initialLayout must not be IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL, IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL, IMAGE_LAYOUT_DEPTH_ATTACHMENT_STENCIL_READ_ONLY_OPTIMAL, or IMAGE_LAYOUT_DEPTH_READ_ONLY_STENCIL_ATTACHMENT_OPTIMAL

If format is a depth/stencil format, initialLayout must not be IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL

If format is a color format, finalLayout must not be IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL, IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL, IMAGE_LAYOUT_DEPTH_ATTACHMENT_STENCIL_READ_ONLY_OPTIMAL, or IMAGE_LAYOUT_DEPTH_READ_ONLY_STENCIL_ATTACHMENT_OPTIMAL

If format is a depth/stencil format, finalLayout must not be IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL

If the separateDepthStencilLayouts feature is not enabled, initialLayout must not be IMAGE_LAYOUT_DEPTH_ATTACHMENT_OPTIMAL, IMAGE_LAYOUT_DEPTH_READ_ONLY_OPTIMAL, IMAGE_LAYOUT_STENCIL_ATTACHMENT_OPTIMAL, or IMAGE_LAYOUT_STENCIL_READ_ONLY_OPTIMAL

If the separateDepthStencilLayouts feature is not enabled, finalLayout must not be IMAGE_LAYOUT_DEPTH_ATTACHMENT_OPTIMAL, IMAGE_LAYOUT_DEPTH_READ_ONLY_OPTIMAL, IMAGE_LAYOUT_STENCIL_ATTACHMENT_OPTIMAL, or IMAGE_LAYOUT_STENCIL_READ_ONLY_OPTIMAL

If format is a color format, initialLayout must not be IMAGE_LAYOUT_DEPTH_ATTACHMENT_OPTIMAL, IMAGE_LAYOUT_DEPTH_READ_ONLY_OPTIMAL, IMAGE_LAYOUT_STENCIL_ATTACHMENT_OPTIMAL, or IMAGE_LAYOUT_STENCIL_READ_ONLY_OPTIMAL

If format is a color format, finalLayout must not be IMAGE_LAYOUT_DEPTH_ATTACHMENT_OPTIMAL, IMAGE_LAYOUT_DEPTH_READ_ONLY_OPTIMAL, IMAGE_LAYOUT_STENCIL_ATTACHMENT_OPTIMAL, or IMAGE_LAYOUT_STENCIL_READ_ONLY_OPTIMAL

If format is a depth/stencil format which includes both depth and stencil aspects, initialLayout must not be IMAGE_LAYOUT_DEPTH_ATTACHMENT_OPTIMAL, IMAGE_LAYOUT_DEPTH_READ_ONLY_OPTIMAL, IMAGE_LAYOUT_STENCIL_ATTACHMENT_OPTIMAL, or IMAGE_LAYOUT_STENCIL_READ_ONLY_OPTIMAL

If format is a depth/stencil format which includes both depth and stencil aspects, finalLayout must not be IMAGE_LAYOUT_DEPTH_ATTACHMENT_OPTIMAL, IMAGE_LAYOUT_DEPTH_READ_ONLY_OPTIMAL, IMAGE_LAYOUT_STENCIL_ATTACHMENT_OPTIMAL, or IMAGE_LAYOUT_STENCIL_READ_ONLY_OPTIMAL

If format is a depth/stencil format which includes only the depth aspect, initialLayout must not be IMAGE_LAYOUT_STENCIL_ATTACHMENT_OPTIMAL or IMAGE_LAYOUT_STENCIL_READ_ONLY_OPTIMAL

If format is a depth/stencil format which includes only the depth aspect, finalLayout must not be IMAGE_LAYOUT_STENCIL_ATTACHMENT_OPTIMAL or IMAGE_LAYOUT_STENCIL_READ_ONLY_OPTIMAL

If format is a depth/stencil format which includes only the stencil aspect, initialLayout must not be IMAGE_LAYOUT_DEPTH_ATTACHMENT_OPTIMAL or IMAGE_LAYOUT_DEPTH_READ_ONLY_OPTIMAL

If format is a depth/stencil format which includes only the stencil aspect, finalLayout must not be IMAGE_LAYOUT_DEPTH_ATTACHMENT_OPTIMAL or IMAGE_LAYOUT_DEPTH_READ_ONLY_OPTIMAL

Valid Usage (Implicit)

flags must be a valid combination of AttachmentDescriptionFlagBits values

format must be a valid Format value

samples must be a valid SampleCountFlagBits value

loadOp must be a valid AttachmentLoadOp value

storeOp must be a valid AttachmentStoreOp value

stencilLoadOp must be a valid AttachmentLoadOp value

stencilStoreOp must be a valid AttachmentStoreOp value

initialLayout must be a valid ImageLayout value

finalLayout must be a valid ImageLayout value

See Also

AttachmentDescriptionFlags, AttachmentLoadOp, AttachmentStoreOp, Format, ImageLayout, RenderPassCreateInfo, SampleCountFlagBits

Constructors

AttachmentDescription

Fields

flags :: AttachmentDescriptionFlags

flags is a bitmask of AttachmentDescriptionFlagBits specifying additional properties of the attachment.

format :: Format

format is a Format value specifying the format of the image view that will be used for the attachment.

samples :: SampleCountFlagBits

samples is the number of samples of the image as defined in SampleCountFlagBits.

loadOp :: AttachmentLoadOp

loadOp is a AttachmentLoadOp value specifying how the contents of color and depth components of the attachment are treated at the beginning of the subpass where it is first used.

storeOp :: AttachmentStoreOp

storeOp is a AttachmentStoreOp value specifying how the contents of color and depth components of the attachment are treated at the end of the subpass where it is last used.

stencilLoadOp :: AttachmentLoadOp

stencilLoadOp is a AttachmentLoadOp value specifying how the contents of stencil components of the attachment are treated at the beginning of the subpass where it is first used.

stencilStoreOp :: AttachmentStoreOp

stencilStoreOp is a AttachmentStoreOp value specifying how the contents of stencil components of the attachment are treated at the end of the last subpass where it is used.

initialLayout :: ImageLayout

initialLayout is the layout the attachment image subresource will be in when a render pass instance begins.

finalLayout :: ImageLayout

finalLayout is the layout the attachment image subresource will be transitioned to when a render pass instance ends.

data AttachmentReference

VkAttachmentReference - Structure specifying an attachment reference

Valid Usage

If attachment is not ATTACHMENT_UNUSED, layout must not be IMAGE_LAYOUT_UNDEFINED, IMAGE_LAYOUT_PREINITIALIZED, IMAGE_LAYOUT_PRESENT_SRC_KHR, IMAGE_LAYOUT_DEPTH_ATTACHMENT_OPTIMAL_KHR, IMAGE_LAYOUT_DEPTH_READ_ONLY_OPTIMAL_KHR, IMAGE_LAYOUT_STENCIL_ATTACHMENT_OPTIMAL_KHR, or IMAGE_LAYOUT_STENCIL_READ_ONLY_OPTIMAL_KHR

Valid Usage (Implicit)

layout must be a valid ImageLayout value

See Also

ImageLayout, RenderPassFragmentDensityMapCreateInfoEXT, SubpassDescription

Constructors

AttachmentReference

Fields

attachment :: Word32

attachment is either an integer value identifying an attachment at the corresponding index in RenderPassCreateInfo::pAttachments, or ATTACHMENT_UNUSED to signify that this attachment is not used.

layout :: ImageLayout

layout is a ImageLayout value specifying the layout the attachment uses during the subpass.

data SubpassDescription

VkSubpassDescription - Structure specifying a subpass description

Description

Each element of the pInputAttachments array corresponds to an input attachment index in a fragment shader, i.e. if a shader declares an image variable decorated with a InputAttachmentIndex value of X, then it uses the attachment provided in pInputAttachments[X]. Input attachments must also be bound to the pipeline in a descriptor set. If the attachment member of any element of pInputAttachments is ATTACHMENT_UNUSED, the application must not read from the corresponding input attachment index. Fragment shaders can use subpass input variables to access the contents of an input attachment at the fragment’s (x, y, layer) framebuffer coordinates. Input attachments must not be used by any subpasses within a renderpass that enables render pass transform.

Each element of the pColorAttachments array corresponds to an output location in the shader, i.e. if the shader declares an output variable decorated with a Location value of X, then it uses the attachment provided in pColorAttachments[X]. If the attachment member of any element of pColorAttachments is ATTACHMENT_UNUSED, writes to the corresponding location by a fragment are discarded.

If pResolveAttachments is not NULL, each of its elements corresponds to a color attachment (the element in pColorAttachments at the same index), and a multisample resolve operation is defined for each attachment. At the end of each subpass, multisample resolve operations read the subpass’s color attachments, and resolve the samples for each pixel within the render area to the same pixel location in the corresponding resolve attachments, unless the resolve attachment index is ATTACHMENT_UNUSED.

Similarly, if SubpassDescriptionDepthStencilResolve::pDepthStencilResolveAttachment is not NULL and does not have the value ATTACHMENT_UNUSED, it corresponds to the depth/stencil attachment in pDepthStencilAttachment, and multisample resolve operations for depth and stencil are defined by SubpassDescriptionDepthStencilResolve::depthResolveMode and SubpassDescriptionDepthStencilResolve::stencilResolveMode, respectively. At the end of each subpass, multisample resolve operations read the subpass’s depth/stencil attachment, and resolve the samples for each pixel to the same pixel location in the corresponding resolve attachment. If SubpassDescriptionDepthStencilResolve::depthResolveMode is RESOLVE_MODE_NONE, then the depth component of the resolve attachment is not written to and its contents are preserved. Similarly, if SubpassDescriptionDepthStencilResolve::stencilResolveMode is RESOLVE_MODE_NONE, then the stencil component of the resolve attachment is not written to and its contents are preserved. SubpassDescriptionDepthStencilResolve::depthResolveMode is ignored if the Format of the pDepthStencilResolveAttachment does not have a depth component. Similarly, SubpassDescriptionDepthStencilResolve::stencilResolveMode is ignored if the Format of the pDepthStencilResolveAttachment does not have a stencil component.

If the image subresource range referenced by the depth/stencil attachment is created with IMAGE_CREATE_SAMPLE_LOCATIONS_COMPATIBLE_DEPTH_BIT_EXT, then the multisample resolve operation uses the sample locations state specified in the sampleLocationsInfo member of the element of the RenderPassSampleLocationsBeginInfoEXT::pPostSubpassSampleLocations for the subpass.

If pDepthStencilAttachment is NULL, or if its attachment index is ATTACHMENT_UNUSED, it indicates that no depth/stencil attachment will be used in the subpass.

The contents of an attachment within the render area become undefined at the start of a subpass S if all of the following conditions are true:

The attachment is used as a color, depth/stencil, or resolve attachment in any subpass in the render pass.

There is a subpass S1 that uses or preserves the attachment, and a subpass dependency from S1 to S.

The attachment is not used or preserved in subpass S.

Once the contents of an attachment become undefined in subpass S, they remain undefined for subpasses in subpass dependency chains starting with subpass S until they are written again. However, they remain valid for subpasses in other subpass dependency chains starting with subpass S1 if those subpasses use or preserve the attachment.

Valid Usage

pipelineBindPoint must be PIPELINE_BIND_POINT_GRAPHICS

colorAttachmentCount must be less than or equal to PhysicalDeviceLimits::maxColorAttachments

If the first use of an attachment in this render pass is as an input attachment, and the attachment is not also used as a color or depth/stencil attachment in the same subpass, then loadOp must not be ATTACHMENT_LOAD_OP_CLEAR

If pResolveAttachments is not NULL, for each resolve attachment that is not ATTACHMENT_UNUSED, the corresponding color attachment must not be ATTACHMENT_UNUSED

If pResolveAttachments is not NULL, for each resolve attachment that is not ATTACHMENT_UNUSED, the corresponding color attachment must not have a sample count of SAMPLE_COUNT_1_BIT

If pResolveAttachments is not NULL, each resolve attachment that is not ATTACHMENT_UNUSED must have a sample count of SAMPLE_COUNT_1_BIT

If pResolveAttachments is not NULL, each resolve attachment that is not ATTACHMENT_UNUSED must have the same Format as its corresponding color attachment

All attachments in pColorAttachments that are not ATTACHMENT_UNUSED must have the same sample count

All attachments in pInputAttachments that are not ATTACHMENT_UNUSED must have formats whose features contain at least one of FORMAT_FEATURE_COLOR_ATTACHMENT_BIT or FORMAT_FEATURE_DEPTH_STENCIL_ATTACHMENT_BIT.

All attachments in pColorAttachments that are not ATTACHMENT_UNUSED must have formats whose features contain FORMAT_FEATURE_COLOR_ATTACHMENT_BIT

All attachments in pResolveAttachments that are not ATTACHMENT_UNUSED must have formats whose features contain FORMAT_FEATURE_COLOR_ATTACHMENT_BIT

If pDepthStencilAttachment is not NULL and the attachment is not ATTACHMENT_UNUSED then it must have a format whose features contain FORMAT_FEATURE_DEPTH_STENCIL_ATTACHMENT_BIT

If the VK_AMD_mixed_attachment_samples extension is enabled, and all attachments in pColorAttachments that are not ATTACHMENT_UNUSED must have a sample count that is smaller than or equal to the sample count of pDepthStencilAttachment if it is not ATTACHMENT_UNUSED

If neither the VK_AMD_mixed_attachment_samples nor the VK_NV_framebuffer_mixed_samples extensions are enabled, and if pDepthStencilAttachment is not ATTACHMENT_UNUSED and any attachments in pColorAttachments are not ATTACHMENT_UNUSED, they must have the same sample count

The attachment member of each element of pPreserveAttachments must not be ATTACHMENT_UNUSED

Each element of pPreserveAttachments must not also be an element of any other member of the subpass description

If any attachment is used by more than one AttachmentReference member, then each use must use the same layout

If flags includes SUBPASS_DESCRIPTION_PER_VIEW_POSITION_X_ONLY_BIT_NVX, it must also include SUBPASS_DESCRIPTION_PER_VIEW_ATTRIBUTES_BIT_NVX.

If the render pass is created with RENDER_PASS_CREATE_TRANSFORM_BIT_QCOM each of the elements of pInputAttachments must be ATTACHMENT_UNUSED.

Valid Usage (Implicit)

flags must be a valid combination of SubpassDescriptionFlagBits values

pipelineBindPoint must be a valid PipelineBindPoint value

If inputAttachmentCount is not 0, pInputAttachments must be a valid pointer to an array of inputAttachmentCount valid AttachmentReference structures

If colorAttachmentCount is not 0, pColorAttachments must be a valid pointer to an array of colorAttachmentCount valid AttachmentReference structures

If colorAttachmentCount is not 0, and pResolveAttachments is not NULL, pResolveAttachments must be a valid pointer to an array of colorAttachmentCount valid AttachmentReference structures

If pDepthStencilAttachment is not NULL, pDepthStencilAttachment must be a valid pointer to a valid AttachmentReference structure

If preserveAttachmentCount is not 0, pPreserveAttachments must be a valid pointer to an array of preserveAttachmentCount uint32_t values

See Also

AttachmentReference, PipelineBindPoint, RenderPassCreateInfo, SubpassDescriptionFlags

Constructors

SubpassDescription

Fields

flags :: SubpassDescriptionFlags

flags is a bitmask of SubpassDescriptionFlagBits specifying usage of the subpass.

pipelineBindPoint :: PipelineBindPoint

pipelineBindPoint is a PipelineBindPoint value specifying the pipeline type supported for this subpass.

inputAttachments :: Vector AttachmentReference

pInputAttachments is a pointer to an array of AttachmentReference structures defining the input attachments for this subpass and their layouts.

colorAttachments :: Vector AttachmentReference

pColorAttachments is a pointer to an array of AttachmentReference structures defining the color attachments for this subpass and their layouts.

resolveAttachments :: Either Word32 (Vector AttachmentReference)

pResolveAttachments is an optional array of colorAttachmentCount AttachmentReference structures defining the resolve attachments for this subpass and their layouts.

depthStencilAttachment :: Maybe AttachmentReference

pDepthStencilAttachment is a pointer to a AttachmentReference structure specifying the depth/stencil attachment for this subpass and its layout.

preserveAttachments :: Vector Word32

pPreserveAttachments is a pointer to an array of preserveAttachmentCount render pass attachment indices identifying attachments that are not used by this subpass, but whose contents must be preserved throughout the subpass.

data SubpassDependency

VkSubpassDependency - Structure specifying a subpass dependency

Description

If srcSubpass is equal to dstSubpass then the SubpassDependency describes a subpass self-dependency, and only constrains the pipeline barriers allowed within a subpass instance. Otherwise, when a render pass instance which includes a subpass dependency is submitted to a queue, it defines a memory dependency between the subpasses identified by srcSubpass and dstSubpass.

If srcSubpass is equal to SUBPASS_EXTERNAL, the first synchronization scope includes commands that occur earlier in submission order than the cmdBeginRenderPass used to begin the render pass instance. Otherwise, the first set of commands includes all commands submitted as part of the subpass instance identified by srcSubpass and any load, store or multisample resolve operations on attachments used in srcSubpass. In either case, the first synchronization scope is limited to operations on the pipeline stages determined by the source stage mask specified by srcStageMask.

If dstSubpass is equal to SUBPASS_EXTERNAL, the second synchronization scope includes commands that occur later in submission order than the cmdEndRenderPass used to end the render pass instance. Otherwise, the second set of commands includes all commands submitted as part of the subpass instance identified by dstSubpass and any load, store or multisample resolve operations on attachments used in dstSubpass. In either case, the second synchronization scope is limited to operations on the pipeline stages determined by the destination stage mask specified by dstStageMask.

The first access scope is limited to access in the pipeline stages determined by the source stage mask specified by srcStageMask. It is also limited to access types in the source access mask specified by srcAccessMask.

The second access scope is limited to access in the pipeline stages determined by the destination stage mask specified by dstStageMask. It is also limited to access types in the destination access mask specified by dstAccessMask.

The availability and visibility operations defined by a subpass dependency affect the execution of image layout transitions within the render pass.

Note

For non-attachment resources, the memory dependency expressed by subpass dependency is nearly identical to that of a MemoryBarrier (with matching srcAccessMask and dstAccessMask parameters) submitted as a part of a cmdPipelineBarrier (with matching srcStageMask and dstStageMask parameters). The only difference being that its scopes are limited to the identified subpasses rather than potentially affecting everything before and after.

For attachments however, subpass dependencies work more like a ImageMemoryBarrier defined similarly to the MemoryBarrier above, the queue family indices set to QUEUE_FAMILY_IGNORED, and layouts as follows:

The equivalent to oldLayout is the attachment’s layout according to the subpass description for srcSubpass.

The equivalent to newLayout is the attachment’s layout according to the subpass description for dstSubpass.

Valid Usage

If the geometry shaders feature is not enabled, srcStageMask must not contain PIPELINE_STAGE_GEOMETRY_SHADER_BIT

If the geometry shaders feature is not enabled, dstStageMask must not contain PIPELINE_STAGE_GEOMETRY_SHADER_BIT

If the tessellation shaders feature is not enabled, srcStageMask must not contain PIPELINE_STAGE_TESSELLATION_CONTROL_SHADER_BIT or PIPELINE_STAGE_TESSELLATION_EVALUATION_SHADER_BIT

If the tessellation shaders feature is not enabled, dstStageMask must not contain PIPELINE_STAGE_TESSELLATION_CONTROL_SHADER_BIT or PIPELINE_STAGE_TESSELLATION_EVALUATION_SHADER_BIT

srcSubpass must be less than or equal to dstSubpass, unless one of them is SUBPASS_EXTERNAL, to avoid cyclic dependencies and ensure a valid execution order

srcSubpass and dstSubpass must not both be equal to SUBPASS_EXTERNAL

If srcSubpass is equal to dstSubpass and not all of the stages in srcStageMask and dstStageMask are framebuffer-space stages, the logically latest pipeline stage in srcStageMask must be logically earlier than or equal to the logically earliest pipeline stage in dstStageMask

Any access flag included in srcAccessMask must be supported by one of the pipeline stages in srcStageMask, as specified in the table of supported access types

Any access flag included in dstAccessMask must be supported by one of the pipeline stages in dstStageMask, as specified in the table of supported access types

If srcSubpass equals dstSubpass, and srcStageMask and dstStageMask both include a framebuffer-space stage, then dependencyFlags must include DEPENDENCY_BY_REGION_BIT

If dependencyFlags includes DEPENDENCY_VIEW_LOCAL_BIT, srcSubpass must not be equal to SUBPASS_EXTERNAL

If dependencyFlags includes DEPENDENCY_VIEW_LOCAL_BIT, dstSubpass must not be equal to SUBPASS_EXTERNAL

If srcSubpass equals dstSubpass and that subpass has more than one bit set in the view mask, then dependencyFlags must include DEPENDENCY_VIEW_LOCAL_BIT

If the mesh shaders feature is not enabled, srcStageMask must not contain PIPELINE_STAGE_MESH_SHADER_BIT_NV

If the task shaders feature is not enabled, srcStageMask must not contain PIPELINE_STAGE_TASK_SHADER_BIT_NV

If the mesh shaders feature is not enabled, dstStageMask must not contain PIPELINE_STAGE_MESH_SHADER_BIT_NV

If the task shaders feature is not enabled, dstStageMask must not contain PIPELINE_STAGE_TASK_SHADER_BIT_NV

Valid Usage (Implicit)

srcStageMask must be a valid combination of PipelineStageFlagBits values

srcStageMask must not be 0

dstStageMask must be a valid combination of PipelineStageFlagBits values

dstStageMask must not be 0

srcAccessMask must be a valid combination of AccessFlagBits values

dstAccessMask must be a valid combination of AccessFlagBits values

dependencyFlags must be a valid combination of DependencyFlagBits values

See Also

AccessFlags, DependencyFlags, PipelineStageFlags, RenderPassCreateInfo

Constructors

SubpassDependency

Fields

srcSubpass :: Word32

srcSubpass is the subpass index of the first subpass in the dependency, or SUBPASS_EXTERNAL.

dstSubpass :: Word32

dstSubpass is the subpass index of the second subpass in the dependency, or SUBPASS_EXTERNAL.

srcStageMask :: PipelineStageFlags

srcStageMask is a bitmask of PipelineStageFlagBits specifying the source stage mask.

dstStageMask :: PipelineStageFlags

dstStageMask is a bitmask of PipelineStageFlagBits specifying the destination stage mask

srcAccessMask :: AccessFlags

srcAccessMask is a bitmask of AccessFlagBits specifying a source access mask.

dstAccessMask :: AccessFlags

dstAccessMask is a bitmask of AccessFlagBits specifying a destination access mask.

dependencyFlags :: DependencyFlags

dependencyFlags is a bitmask of DependencyFlagBits.

data RenderPassCreateInfo (es :: [Type])

VkRenderPassCreateInfo - Structure specifying parameters of a newly created render pass

Description

Note

Care should be taken to avoid a data race here; if any subpasses access attachments with overlapping memory locations, and one of those accesses is a write, a subpass dependency needs to be included between them.

Valid Usage

If the attachment member of any element of pInputAttachments, pColorAttachments, pResolveAttachments or pDepthStencilAttachment, or any element of pPreserveAttachments in any element of pSubpasses is not ATTACHMENT_UNUSED, it must be less than attachmentCount

For any member of pAttachments with a loadOp equal to ATTACHMENT_LOAD_OP_CLEAR, the first use of that attachment must not specify a layout equal to IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL or IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL.

For any member of pAttachments with a stencilLoadOp equal to ATTACHMENT_LOAD_OP_CLEAR, the first use of that attachment must not specify a layout equal to IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL or IMAGE_LAYOUT_DEPTH_STENCIL_READ_ONLY_OPTIMAL.

For any member of pAttachments with a loadOp equal to ATTACHMENT_LOAD_OP_CLEAR, the first use of that attachment must not specify a layout equal to IMAGE_LAYOUT_DEPTH_READ_ONLY_STENCIL_ATTACHMENT_OPTIMAL.

For any member of pAttachments with a stencilLoadOp equal to ATTACHMENT_LOAD_OP_CLEAR, the first use of that attachment must not specify a layout equal to IMAGE_LAYOUT_DEPTH_ATTACHMENT_STENCIL_READ_ONLY_OPTIMAL.

If the pNext chain includes a RenderPassInputAttachmentAspectCreateInfo structure, the subpass member of each element of its pAspectReferences member must be less than subpassCount

If the pNext chain includes a RenderPassInputAttachmentAspectCreateInfo structure, the inputAttachmentIndex member of each element of its pAspectReferences member must be less than the value of inputAttachmentCount in the member of pSubpasses identified by its subpass member

If the pNext chain includes a RenderPassInputAttachmentAspectCreateInfo structure, for any element of the pInputAttachments member of any element of pSubpasses where the attachment member is not ATTACHMENT_UNUSED, the aspectMask member of the corresponding element of RenderPassInputAttachmentAspectCreateInfo::pAspectReferences must only include aspects that are present in images of the format specified by the element of pAttachments at attachment

If the pNext chain includes a RenderPassMultiviewCreateInfo structure, and its subpassCount member is not zero, that member must be equal to the value of subpassCount

If the pNext chain includes a RenderPassMultiviewCreateInfo structure, if its dependencyCount member is not zero, it must be equal to dependencyCount

If the pNext chain includes a RenderPassMultiviewCreateInfo structure, for each non-zero element of pViewOffsets, the srcSubpass and dstSubpass members of pDependencies at the same index must not be equal

If the pNext chain includes a RenderPassMultiviewCreateInfo structure, for any element of pDependencies with a dependencyFlags member that does not include DEPENDENCY_VIEW_LOCAL_BIT, the corresponding element of the pViewOffsets member of that RenderPassMultiviewCreateInfo instance must be 0

If the pNext chain includes a RenderPassMultiviewCreateInfo structure, elements of its pViewMasks member must either all be 0, or all not be 0

If the pNext chain includes a RenderPassMultiviewCreateInfo structure, and each element of its pViewMasks member is 0, the dependencyFlags member of each element of pDependencies must not include DEPENDENCY_VIEW_LOCAL_BIT

If the pNext chain includes a RenderPassMultiviewCreateInfo structure, and each element of its pViewMasks member is 0, correlatedViewMaskCount must be 0

If the pNext chain includes a RenderPassMultiviewCreateInfo structure, each element of its pViewMask member must not have a bit set at an index greater than or equal to PhysicalDeviceLimits::maxFramebufferLayers

For any element of pDependencies, if the srcSubpass is not SUBPASS_EXTERNAL, all stage flags included in the srcStageMask member of that dependency must be a pipeline stage supported by the pipeline identified by the pipelineBindPoint member of the source subpass

For any element of pDependencies, if the dstSubpass is not SUBPASS_EXTERNAL, all stage flags included in the dstStageMask member of that dependency must be a pipeline stage supported by the pipeline identified by the pipelineBindPoint member of the destination subpass

The srcSubpass member of each element of pDependencies must be less than subpassCount

The dstSubpass member of each element of pDependencies must be less than subpassCount

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of RenderPassFragmentDensityMapCreateInfoEXT, RenderPassInputAttachmentAspectCreateInfo, or RenderPassMultiviewCreateInfo

The sType value of each struct in the pNext chain must be unique

flags must be a valid combination of RenderPassCreateFlagBits values

If attachmentCount is not 0, pAttachments must be a valid pointer to an array of attachmentCount valid AttachmentDescription structures

pSubpasses must be a valid pointer to an array of subpassCount valid SubpassDescription structures

If dependencyCount is not 0, pDependencies must be a valid pointer to an array of dependencyCount valid SubpassDependency structures

subpassCount must be greater than 0

See Also

AttachmentDescription, RenderPassCreateFlags, StructureType, SubpassDependency, SubpassDescription, createRenderPass

Constructors

RenderPassCreateInfo

Fields

next :: Chain es

pNext is NULL or a pointer to an extension-specific structure.

flags :: RenderPassCreateFlags

flags is a bitmask of RenderPassCreateFlagBits

attachments :: Vector AttachmentDescription

pAttachments is a pointer to an array of attachmentCount AttachmentDescription structures describing the attachments used by the render pass.

subpasses :: Vector SubpassDescription

pSubpasses is a pointer to an array of subpassCount SubpassDescription structures describing each subpass.

dependencies :: Vector SubpassDependency

pDependencies is a pointer to an array of dependencyCount SubpassDependency structures describing dependencies between pairs of subpasses.

data FramebufferCreateInfo (es :: [Type])

VkFramebufferCreateInfo - Structure specifying parameters of a newly created framebuffer

Description

Applications must ensure that all accesses to memory that backs image subresources used as attachments in a given renderpass instance either happen-before the load operations for those attachments, or happen-after the store operations for those attachments.

For depth/stencil attachments, each aspect can be used separately as attachments and non-attachments as long as the non-attachment accesses are also via an image subresource in either the IMAGE_LAYOUT_DEPTH_READ_ONLY_STENCIL_ATTACHMENT_OPTIMAL layout or the IMAGE_LAYOUT_DEPTH_ATTACHMENT_STENCIL_READ_ONLY_OPTIMAL layout, and the attachment resource uses whichever of those two layouts the image accesses do not. Use of non-attachment aspects in this case is only well defined if the attachment is used in the subpass where the non-attachment access is being made, or the layout of the image subresource is constant throughout the entire render pass instance, including the initialLayout and finalLayout.

Note

These restrictions mean that the render pass has full knowledge of all uses of all of the attachments, so that the implementation is able to make correct decisions about when and how to perform layout transitions, when to overlap execution of subpasses, etc.

It is legal for a subpass to use no color or depth/stencil attachments, and rather use shader side effects such as image stores and atomics to produce an output. In this case, the subpass continues to use the width, height, and layers of the framebuffer to define the dimensions of the rendering area, and the rasterizationSamples from each pipeline’s PipelineMultisampleStateCreateInfo to define the number of samples used in rasterization; however, if PhysicalDeviceFeatures::variableMultisampleRate is FALSE, then all pipelines to be bound with a given zero-attachment subpass must have the same value for PipelineMultisampleStateCreateInfo::rasterizationSamples.

Valid Usage

attachmentCount must be equal to the attachment count specified in renderPass

If flags does not include FRAMEBUFFER_CREATE_IMAGELESS_BIT, and attachmentCount is not 0, pAttachments must be a valid pointer to an array of attachmentCount valid ImageView handles

If flags does not include FRAMEBUFFER_CREATE_IMAGELESS_BIT, each element of pAttachments that is used as a color attachment or resolve attachment by renderPass must have been created with a usage value including IMAGE_USAGE_COLOR_ATTACHMENT_BIT

If flags does not include FRAMEBUFFER_CREATE_IMAGELESS_BIT, each element of pAttachments that is used as a depth/stencil attachment by renderPass must have been created with a usage value including IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT

If flags does not include FRAMEBUFFER_CREATE_IMAGELESS_BIT, each element of pAttachments that is used as a depth/stencil resolve attachment by renderPass must have been created with a usage value including IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT

If flags does not include FRAMEBUFFER_CREATE_IMAGELESS_BIT, each element of pAttachments that is used as an input attachment by renderPass must have been created with a usage value including IMAGE_USAGE_INPUT_ATTACHMENT_BIT

Each element of pAttachments that is used as a fragment density map attachment by renderPass must not have been created with a flags value including IMAGE_CREATE_SUBSAMPLED_BIT_EXT.

If renderPass has a fragment density map attachment and non-subsample image feature is not enabled, each element of pAttachments must have been created with a flags value including IMAGE_CREATE_SUBSAMPLED_BIT_EXT unless that element is the fragment density map attachment.

If flags does not include FRAMEBUFFER_CREATE_IMAGELESS_BIT, each element of pAttachments must have been created with a Format value that matches the Format specified by the corresponding AttachmentDescription in renderPass

If flags does not include FRAMEBUFFER_CREATE_IMAGELESS_BIT, each element of pAttachments must have been created with a samples value that matches the samples value specified by the corresponding AttachmentDescription in renderPass

If flags does not include FRAMEBUFFER_CREATE_IMAGELESS_BIT, each element of pAttachments must have dimensions at least as large as the corresponding framebuffer dimension except for any element that is referenced by fragmentDensityMapAttachment

If renderPass was specified with non-zero view masks, each element of pAttachments that is not referenced by fragmentDensityMapAttachment must have a layerCount greater than the index of the most significant bit set in any of those view masks

If renderPass was specified with non-zero view masks, each element of pAttachments that is referenced by fragmentDensityMapAttachment must have a layerCount equal to 1 or greater than the index of the most significant bit set in any of those view masks

If renderPass was not specified with non-zero view masks, each element of pAttachments that is referenced by fragmentDensityMapAttachment must have a layerCount equal to 1

If flags does not include FRAMEBUFFER_CREATE_IMAGELESS_BIT, an element of pAttachments that is referenced by fragmentDensityMapAttachment must have a width at least as large as \(\lceil{\frac{width}{maxFragmentDensityTexelSize_{width}}}\rceil\)

If flags does not include FRAMEBUFFER_CREATE_IMAGELESS_BIT, an element of pAttachments that is referenced by fragmentDensityMapAttachment must have a height at least as large as \(\lceil{\frac{height}{maxFragmentDensityTexelSize_{height}}}\rceil\)

If flags does not include FRAMEBUFFER_CREATE_IMAGELESS_BIT, each element of pAttachments must only specify a single mip level

If flags does not include FRAMEBUFFER_CREATE_IMAGELESS_BIT, each element of pAttachments must have been created with the identity swizzle

width must be greater than 0.

width must be less than or equal to PhysicalDeviceLimits::maxFramebufferWidth

height must be greater than 0.

height must be less than or equal to PhysicalDeviceLimits::maxFramebufferHeight

layers must be greater than 0.

layers must be less than or equal to PhysicalDeviceLimits::maxFramebufferLayers

If renderPass was specified with non-zero view masks, layers must be 1

If flags does not include FRAMEBUFFER_CREATE_IMAGELESS_BIT, each element of pAttachments that is a 2D or 2D array image view taken from a 3D image must not be a depth/stencil format

If flags does not include FRAMEBUFFER_CREATE_IMAGELESS_BIT, and attachmentCount is not 0, pAttachments must be a valid pointer to an array of attachmentCount valid ImageView handles

If the imageless framebuffer feature is not enabled, flags must not include FRAMEBUFFER_CREATE_IMAGELESS_BIT

If flags includes FRAMEBUFFER_CREATE_IMAGELESS_BIT, the pNext chain must include a FramebufferAttachmentsCreateInfo structure

If flags includes FRAMEBUFFER_CREATE_IMAGELESS_BIT, the attachmentImageInfoCount member of a FramebufferAttachmentsCreateInfo structure included in the pNext chain must be equal to either zero or attachmentCount

If flags includes FRAMEBUFFER_CREATE_IMAGELESS_BIT, the width member of any element of the pAttachmentImageInfos member of a FramebufferAttachmentsCreateInfo structure included in the pNext chain must be greater than or equal to width, except for any element that is referenced by RenderPassFragmentDensityMapCreateInfoEXT::fragmentDensityMapAttachment in renderPass

If flags includes FRAMEBUFFER_CREATE_IMAGELESS_BIT, the height member of any element of the pAttachmentImageInfos member of a FramebufferAttachmentsCreateInfo structure included in the pNext chain must be greater than or equal to height, except for any element that is referenced by RenderPassFragmentDensityMapCreateInfoEXT::fragmentDensityMapAttachment in renderPass

If flags includes FRAMEBUFFER_CREATE_IMAGELESS_BIT, the width member of any element of the pAttachmentImageInfos member of a FramebufferAttachmentsCreateInfo structure included in the pNext chain that is referenced by RenderPassFragmentDensityMapCreateInfoEXT::fragmentDensityMapAttachment in renderPass must be greater than or equal to \(\lceil{\frac{width}{maxFragmentDensityTexelSize_{width}}}\rceil\)

If flags includes FRAMEBUFFER_CREATE_IMAGELESS_BIT, the height member of any element of the pAttachmentImageInfos member of a FramebufferAttachmentsCreateInfo structure included in the pNext chain that is referenced by RenderPassFragmentDensityMapCreateInfoEXT::fragmentDensityMapAttachment in renderPass must be greater than or equal to \(\lceil{\frac{height}{maxFragmentDensityTexelSize_{height}}}\rceil\)

If multiview is enabled for renderPass, and flags includes FRAMEBUFFER_CREATE_IMAGELESS_BIT, the layerCount member of any element of the pAttachmentImageInfos member of a FramebufferAttachmentsCreateInfo structure included in the pNext chain must be greater than the maximum bit index set in the view mask in the subpasses in which it is used in renderPass

If multiview is not enabled for renderPass, and flags includes FRAMEBUFFER_CREATE_IMAGELESS_BIT, the layerCount member of any element of the pAttachmentImageInfos member of a FramebufferAttachmentsCreateInfo structure included in the pNext chain must be greater than or equal to layers

If flags includes FRAMEBUFFER_CREATE_IMAGELESS_BIT, the usage member of any element of the pAttachmentImageInfos member of a FramebufferAttachmentsCreateInfo structure included in the pNext chain that refers to an attachment used as a color attachment or resolve attachment by renderPass must include IMAGE_USAGE_COLOR_ATTACHMENT_BIT

If flags includes FRAMEBUFFER_CREATE_IMAGELESS_BIT, the usage member of any element of the pAttachmentImageInfos member of a FramebufferAttachmentsCreateInfo structure included in the pNext chain that refers to an attachment used as a depth/stencil attachment by renderPass must include IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT

If flags includes FRAMEBUFFER_CREATE_IMAGELESS_BIT, the usage member of any element of the pAttachmentImageInfos member of a FramebufferAttachmentsCreateInfo structure included in the pNext chain that refers to an attachment used as a depth/stencil resolve attachment by renderPass must include IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT

If flags includes FRAMEBUFFER_CREATE_IMAGELESS_BIT, the usage member of any element of the pAttachmentImageInfos member of a FramebufferAttachmentsCreateInfo structure included in the pNext chain that refers to an attachment used as an input attachment by renderPass must include IMAGE_USAGE_INPUT_ATTACHMENT_BIT

If flags includes FRAMEBUFFER_CREATE_IMAGELESS_BIT, at least one element of the pViewFormats member of any element of the pAttachmentImageInfos member of a FramebufferAttachmentsCreateInfo structure included in the pNext chain must be equal to the corresponding value of AttachmentDescription::format used to create renderPass

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO

pNext must be NULL or a pointer to a valid instance of FramebufferAttachmentsCreateInfo

The sType value of each struct in the pNext chain must be unique

flags must be a valid combination of FramebufferCreateFlagBits values

renderPass must be a valid RenderPass handle

Both of renderPass, and the elements of pAttachments that are valid handles of non-ignored parameters must have been created, allocated, or retrieved from the same Device

See Also

FramebufferCreateFlags, ImageView, RenderPass, StructureType, createFramebuffer

Constructors

FramebufferCreateInfo

Fields

next :: Chain es

pNext is NULL or a pointer to an extension-specific structure.

flags :: FramebufferCreateFlags

flags is a bitmask of FramebufferCreateFlagBits

renderPass :: RenderPass

renderPass is a render pass defining what render passes the framebuffer will be compatible with. See Render Pass Compatibility for details.

attachments :: Vector ImageView

pAttachments is a pointer to an array of ImageView handles, each of which will be used as the corresponding attachment in a render pass instance. If flags includes FRAMEBUFFER_CREATE_IMAGELESS_BIT, this parameter is ignored.

width :: Word32

width, height and layers define the dimensions of the framebuffer. If the render pass uses multiview, then layers must be one and each attachment requires a number of layers that is greater than the maximum bit index set in the view mask in the subpasses in which it is used.

height :: Word32

layers :: Word32