Graphics.Vulkan.Core10.PipelineCache

Documentation

createPipelineCache :: Device -> PipelineCacheCreateInfo -> ("allocator" ::: Maybe AllocationCallbacks) -> IO PipelineCache

vkCreatePipelineCache - Creates a new pipeline cache

Parameters

Device is the logical device that creates the pipeline cache object.

pCreateInfo is a pointer to a PipelineCacheCreateInfo structure containing initial parameters for the pipeline cache object.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pPipelineCache is a pointer to a PipelineCache handle in which the resulting pipeline cache object is returned.

Description

Note

Applications can track and manage the total host memory size of a pipeline cache object using the pAllocator. Applications can limit the amount of data retrieved from a pipeline cache object in getPipelineCacheData. Implementations should not internally limit the total number of entries added to a pipeline cache object or the total host memory consumed.

Once created, a pipeline cache can be passed to the createGraphicsPipelines and createComputePipelines commands. If the pipeline cache passed into these commands is not NULL_HANDLE, the implementation will query it for possible reuse opportunities and update it with new content. The use of the pipeline cache object in these commands is internally synchronized, and the same pipeline cache object can be used in multiple threads simultaneously.

Note

Implementations should make every effort to limit any critical sections to the actual accesses to the cache, which is expected to be significantly shorter than the duration of the createGraphicsPipelines and createComputePipelines commands.

Valid Usage (Implicit)

Device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid PipelineCacheCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pPipelineCache must be a valid pointer to a PipelineCache handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, Device, PipelineCache, PipelineCacheCreateInfo

withPipelineCache :: Device -> PipelineCacheCreateInfo -> Maybe AllocationCallbacks -> (PipelineCache -> IO r) -> IO r

A safe wrapper for createPipelineCache and destroyPipelineCache using bracket

The allocated value must not be returned from the provided computation

destroyPipelineCache :: Device -> PipelineCache -> ("allocator" ::: Maybe AllocationCallbacks) -> IO ()

vkDestroyPipelineCache - Destroy a pipeline cache object

Parameters

Device is the logical device that destroys the pipeline cache object.

PipelineCache is the handle of the pipeline cache to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

If AllocationCallbacks were provided when PipelineCache was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when PipelineCache was created, pAllocator must be NULL

Valid Usage (Implicit)

Device must be a valid Device handle

If PipelineCache is not NULL_HANDLE, PipelineCache must be a valid PipelineCache handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If PipelineCache is a valid handle, it must have been created, allocated, or retrieved from Device

Host Synchronization

Host access to PipelineCache must be externally synchronized

See Also

AllocationCallbacks, Device, PipelineCache

getPipelineCacheData :: Device -> PipelineCache -> IO (Result, "data" ::: ByteString)

vkGetPipelineCacheData - Get the data store from a pipeline cache

Parameters

Device is the logical device that owns the pipeline cache.

PipelineCache is the pipeline cache to retrieve data from.

pDataSize is a pointer to a size_t value related to the amount of data in the pipeline cache, as described below.

pData is either NULL or a pointer to a buffer.

Description

If pData is NULL, then the maximum size of the data that can be retrieved from the pipeline cache, in bytes, is returned in pDataSize. Otherwise, pDataSize must point to a variable set by the user to the size of the buffer, in bytes, pointed to by pData, and on return the variable is overwritten with the amount of data actually written to pData.

If pDataSize is less than the maximum size that can be retrieved by the pipeline cache, at most pDataSize bytes will be written to pData, and getPipelineCacheData will return INCOMPLETE. Any data written to pData is valid and can be provided as the pInitialData member of the PipelineCacheCreateInfo structure passed to createPipelineCache.

Two calls to getPipelineCacheData with the same parameters must retrieve the same data unless a command that modifies the contents of the cache is called between them.

Applications can store the data retrieved from the pipeline cache, and use these data, possibly in a future run of the application, to populate new pipeline cache objects. The results of pipeline compiles, however, may depend on the vendor ID, device ID, driver version, and other details of the device. To enable applications to detect when previously retrieved data is incompatible with the device, the initial bytes written to pData must be a header consisting of the following members:

Offset 	 Size 	 Meaning

0 	 4 	 length in bytes of the entire pipeline cache header written as a stream of bytes, with the least significant byte first

4 	 4 	 a PipelineCacheHeaderVersion value written as a stream of bytes, with the least significant byte first

8 	 4 	 a vendor ID equal to PhysicalDeviceProperties::vendorID written as a stream of bytes, with the least significant byte first

12 	 4 	 a device ID equal to PhysicalDeviceProperties::deviceID written as a stream of bytes, with the least significant byte first

16 	 UUID_SIZE 	 a pipeline cache ID equal to PhysicalDeviceProperties::pipelineCacheUUID

Layout for pipeline cache header version PIPELINE_CACHE_HEADER_VERSION_ONE

The first four bytes encode the length of the entire pipeline cache header, in bytes. This value includes all fields in the header including the pipeline cache version field and the size of the length field.

The next four bytes encode the pipeline cache version, as described for PipelineCacheHeaderVersion. A consumer of the pipeline cache should use the cache version to interpret the remainder of the cache header.

If pDataSize is less than what is necessary to store this header, nothing will be written to pData and zero will be written to pDataSize.

Valid Usage (Implicit)

Device must be a valid Device handle

PipelineCache must be a valid PipelineCache handle

pDataSize must be a valid pointer to a size_t value

If the value referenced by pDataSize is not 0, and pData is not NULL, pData must be a valid pointer to an array of pDataSize bytes

PipelineCache must have been created, allocated, or retrieved from Device

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

Device, PipelineCache

mergePipelineCaches :: Device -> ("dstCache" ::: PipelineCache) -> ("srcCaches" ::: Vector PipelineCache) -> IO ()

vkMergePipelineCaches - Combine the data stores of pipeline caches

Parameters

Device is the logical device that owns the pipeline cache objects.

dstCache is the handle of the pipeline cache to merge results into.

srcCacheCount is the length of the pSrcCaches array.

pSrcCaches is a pointer to an array of pipeline cache handles, which will be merged into dstCache. The previous contents of dstCache are included after the merge.

Description

Note

The details of the merge operation are implementation dependent, but implementations should merge the contents of the specified pipelines and prune duplicate entries.

Valid Usage

dstCache must not appear in the list of source caches

Valid Usage (Implicit)

Device must be a valid Device handle

dstCache must be a valid PipelineCache handle

pSrcCaches must be a valid pointer to an array of srcCacheCount valid PipelineCache handles

srcCacheCount must be greater than 0

dstCache must have been created, allocated, or retrieved from Device

Each element of pSrcCaches must have been created, allocated, or retrieved from Device

Host Synchronization

Host access to dstCache must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

Device, PipelineCache

data PipelineCacheCreateInfo

VkPipelineCacheCreateInfo - Structure specifying parameters of a newly created pipeline cache

Valid Usage

If initialDataSize is not 0, it must be equal to the size of pInitialData, as returned by getPipelineCacheData when pInitialData was originally retrieved

If initialDataSize is not 0, pInitialData must have been retrieved from a previous call to getPipelineCacheData

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PIPELINE_CACHE_CREATE_INFO

pNext must be NULL

Flags must be 0

If initialDataSize is not 0, pInitialData must be a valid pointer to an array of initialDataSize bytes

See Also

PipelineCacheCreateFlags, StructureType, createPipelineCache

Constructors

PipelineCacheCreateInfo

Fields

flags :: PipelineCacheCreateFlags

Flags is reserved for future use.

initialDataSize :: Word64

initialDataSize is the number of bytes in pInitialData. If initialDataSize is zero, the pipeline cache will initially be empty.

initialData :: Ptr ()

pInitialData is a pointer to previously retrieved pipeline cache data. If the pipeline cache data is incompatible (as defined below) with the device, the pipeline cache will be initially empty. If initialDataSize is zero, pInitialData is ignored.