Graphics.Vulkan.Core10.QueueSemaphore

Documentation

createSemaphore :: forall a io. (PokeChain a, MonadIO io) => Device -> SemaphoreCreateInfo a -> ("allocator" ::: Maybe AllocationCallbacks) -> io Semaphore

vkCreateSemaphore - Create a new queue semaphore object

Parameters

device is the logical device that creates the semaphore.

pCreateInfo is a pointer to a SemaphoreCreateInfo structure containing information about how the semaphore is to be created.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pSemaphore is a pointer to a handle in which the resulting semaphore object is returned.

Valid Usage (Implicit)

device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid SemaphoreCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pSemaphore must be a valid pointer to a Semaphore handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, Device, Semaphore, SemaphoreCreateInfo

withSemaphore :: forall a r. PokeChain a => Device -> SemaphoreCreateInfo a -> Maybe AllocationCallbacks -> (Semaphore -> IO r) -> IO r

A safe wrapper for createSemaphore and destroySemaphore using bracket

The allocated value must not be returned from the provided computation

destroySemaphore :: forall io. MonadIO io => Device -> Semaphore -> ("allocator" ::: Maybe AllocationCallbacks) -> io ()

vkDestroySemaphore - Destroy a semaphore object

Parameters

device is the logical device that destroys the semaphore.

semaphore is the handle of the semaphore to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted batches that refer to semaphore must have completed execution

If AllocationCallbacks were provided when semaphore was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when semaphore was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid Device handle

If semaphore is not NULL_HANDLE, semaphore must be a valid Semaphore handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If semaphore is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to semaphore must be externally synchronized

See Also

AllocationCallbacks, Device, Semaphore

data SemaphoreCreateInfo (es :: [Type])

VkSemaphoreCreateInfo - Structure specifying parameters of a newly created semaphore

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_SEMAPHORE_CREATE_INFO

Each pNext member of any structure (including this one) in the pNext chain must be either NULL or a pointer to a valid instance of ExportSemaphoreCreateInfo, ExportSemaphoreWin32HandleInfoKHR, or SemaphoreTypeCreateInfo

The sType value of each struct in the pNext chain must be unique

flags must be 0

See Also

SemaphoreCreateFlags, StructureType, createSemaphore

Constructors

SemaphoreCreateInfo

Fields

next :: Chain es

pNext is NULL or a pointer to an extension-specific structure.

flags :: SemaphoreCreateFlags

flags is reserved for future use.