Graphics.Vulkan.Extensions.VK_AMD_shader_info

Documentation

getShaderInfoAMD :: Device -> Pipeline -> ShaderStageFlagBits -> ShaderInfoTypeAMD -> IO (Result, "info" ::: ByteString)

vkGetShaderInfoAMD - Get information about a shader in a pipeline

Parameters

Device is the device that created Pipeline.

Pipeline is the target of the query.

shaderStage identifies the particular shader within the pipeline about which information is being queried.

infoType describes what kind of information is being queried.

pInfoSize is a pointer to a value related to the amount of data the query returns, as described below.

pInfo is either NULL or a pointer to a buffer.

Description

If pInfo is NULL, then the maximum size of the information that can be retrieved about the shader, in bytes, is returned in pInfoSize. Otherwise, pInfoSize must point to a variable set by the user to the size of the buffer, in bytes, pointed to by pInfo, and on return the variable is overwritten with the amount of data actually written to pInfo.

If pInfoSize is less than the maximum size that can be retrieved by the pipeline cache, then at most pInfoSize bytes will be written to pInfo, and getShaderInfoAMD will return INCOMPLETE.

Not all information is available for every shader and implementations may not support all kinds of information for any shader. When a certain type of information is unavailable, the function returns ERROR_FEATURE_NOT_PRESENT.

If information is successfully and fully queried, the function will return SUCCESS.

For infoType SHADER_INFO_TYPE_STATISTICS_AMD, a ShaderStatisticsInfoAMD structure will be written to the buffer pointed to by pInfo. This structure will be populated with statistics regarding the physical device resources used by that shader along with other miscellaneous information and is described in further detail below.

For infoType SHADER_INFO_TYPE_DISASSEMBLY_AMD, pInfo is a pointer to a UTF-8 null-terminated string containing human-readable disassembly. The exact formatting and contents of the disassembly string are vendor-specific.

The formatting and contents of all other types of information, including infoType SHADER_INFO_TYPE_BINARY_AMD, are left to the vendor and are not further specified by this extension.

Valid Usage (Implicit)

Device must be a valid Device handle

Pipeline must be a valid Pipeline handle

shaderStage must be a valid ShaderStageFlagBits value

infoType must be a valid ShaderInfoTypeAMD value

pInfoSize must be a valid pointer to a size_t value

If the value referenced by pInfoSize is not 0, and pInfo is not NULL, pInfo must be a valid pointer to an array of pInfoSize bytes

Pipeline must have been created, allocated, or retrieved from Device

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_FEATURE_NOT_PRESENT

ERROR_OUT_OF_HOST_MEMORY

See Also

Device, Pipeline, ShaderInfoTypeAMD, ShaderStageFlagBits

data ShaderResourceUsageAMD

VkShaderResourceUsageAMD - Resource usage information about a particular shader within a pipeline

See Also

ShaderStatisticsInfoAMD

Constructors

ShaderResourceUsageAMD

Fields

numUsedVgprs :: Word32

numUsedVgprs is the number of vector instruction general-purpose registers used by this shader.

numUsedSgprs :: Word32

numUsedSgprs is the number of scalar instruction general-purpose registers used by this shader.

ldsSizePerLocalWorkGroup :: Word32

ldsSizePerLocalWorkGroup is the maximum local data store size per work group in bytes.

ldsUsageSizeInBytes :: Word64

ldsUsageSizeInBytes is the LDS usage size in bytes per work group by this shader.

scratchMemUsageInBytes :: Word64

scratchMemUsageInBytes is the scratch memory usage in bytes by this shader.

data ShaderStatisticsInfoAMD

VkShaderStatisticsInfoAMD - Statistical information about a particular shader within a pipeline

Description

Some implementations may merge multiple logical shader stages together in a single shader. In such cases, shaderStageMask will contain a bitmask of all of the stages that are active within that shader. Consequently, if specifying those stages as input to getShaderInfoAMD, the same output information may be returned for all such shader stage queries.

The number of available VGPRs and SGPRs (numAvailableVgprs and numAvailableSgprs respectively) are the shader-addressable subset of physical registers that is given as a limit to the compiler for register assignment. These values may further be limited by implementations due to performance optimizations where register pressure is a bottleneck.

See Also

ShaderResourceUsageAMD, ShaderStageFlags

Constructors

ShaderStatisticsInfoAMD

Fields

shaderStageMask :: ShaderStageFlags

shaderStageMask are the combination of logical shader stages contained within this shader.

resourceUsage :: ShaderResourceUsageAMD

resourceUsage is a ShaderResourceUsageAMD structure describing internal physical device resources used by this shader.

numPhysicalVgprs :: Word32

numPhysicalVgprs is the maximum number of vector instruction general-purpose registers (VGPRs) available to the physical device.

numPhysicalSgprs :: Word32

numPhysicalSgprs is the maximum number of scalar instruction general-purpose registers (SGPRs) available to the physical device.

numAvailableVgprs :: Word32

numAvailableVgprs is the maximum limit of VGPRs made available to the shader compiler.

numAvailableSgprs :: Word32

numAvailableSgprs is the maximum limit of SGPRs made available to the shader compiler.

computeWorkGroupSize :: (Word32, Word32, Word32)

computeWorkGroupSize is the local workgroup size of this shader in { X, Y, Z } dimensions.

newtype ShaderInfoTypeAMD

VkShaderInfoTypeAMD - Enum specifying which type of shader info to query

See Also

getShaderInfoAMD

Constructors

ShaderInfoTypeAMD Int32

Bundled Patterns

pattern SHADER_INFO_TYPE_STATISTICS_AMD :: ShaderInfoTypeAMD
SHADER_INFO_TYPE_STATISTICS_AMD specifies that device resources used by a shader will be queried.

pattern SHADER_INFO_TYPE_BINARY_AMD :: ShaderInfoTypeAMD
SHADER_INFO_TYPE_BINARY_AMD specifies that implementation-specific information will be queried.

pattern SHADER_INFO_TYPE_DISASSEMBLY_AMD :: ShaderInfoTypeAMD
SHADER_INFO_TYPE_DISASSEMBLY_AMD specifies that human-readable dissassembly of a shader.

type AMD_SHADER_INFO_SPEC_VERSION = 1

pattern AMD_SHADER_INFO_SPEC_VERSION :: forall a. Integral a => a

type AMD_SHADER_INFO_EXTENSION_NAME = "VK_AMD_shader_info"

pattern AMD_SHADER_INFO_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a