Graphics.Vulkan.Extensions.VK_EXT_debug_marker

Documentation

withCStructDebugMarkerMarkerInfoEXT :: DebugMarkerMarkerInfoEXT -> (VkDebugMarkerMarkerInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDebugMarkerMarkerInfoEXT and marshal a DebugMarkerMarkerInfoEXT into it. The VkDebugMarkerMarkerInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDebugMarkerMarkerInfoEXT :: VkDebugMarkerMarkerInfoEXT -> IO DebugMarkerMarkerInfoEXT

A function to read a VkDebugMarkerMarkerInfoEXT and all additional structures in the pointer chain into a DebugMarkerMarkerInfoEXT.

data DebugMarkerMarkerInfoEXT

VkDebugMarkerMarkerInfoEXT - Specify parameters of a command buffer marker region

Valid Usage (Implicit)

See Also

VkStructureType, vkCmdDebugMarkerBeginEXT, vkCmdDebugMarkerInsertEXT

Constructors

DebugMarkerMarkerInfoEXT

Fields

next :: Maybe SomeVkStruct

markerName :: ByteString

color :: (CFloat, CFloat, CFloat, CFloat)

withCStructDebugMarkerObjectNameInfoEXT :: DebugMarkerObjectNameInfoEXT -> (VkDebugMarkerObjectNameInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDebugMarkerObjectNameInfoEXT and marshal a DebugMarkerObjectNameInfoEXT into it. The VkDebugMarkerObjectNameInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDebugMarkerObjectNameInfoEXT :: VkDebugMarkerObjectNameInfoEXT -> IO DebugMarkerObjectNameInfoEXT

A function to read a VkDebugMarkerObjectNameInfoEXT and all additional structures in the pointer chain into a DebugMarkerObjectNameInfoEXT.

data DebugMarkerObjectNameInfoEXT

VkDebugMarkerObjectNameInfoEXT - Specify parameters of a name to give to an object

Description

Applications may change the name associated with an object simply by calling vkDebugMarkerSetObjectNameEXT again with a new string. To remove a previously set name, pObjectName should be set to an empty string.

Valid Usage (Implicit)

See Also

VkDebugReportObjectTypeEXT, VkStructureType, vkDebugMarkerSetObjectNameEXT

Constructors

DebugMarkerObjectNameInfoEXT

Fields

next :: Maybe SomeVkStruct

objectType :: DebugReportObjectTypeEXT

object :: Word64

objectName :: ByteString

withCStructDebugMarkerObjectTagInfoEXT :: DebugMarkerObjectTagInfoEXT -> (VkDebugMarkerObjectTagInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkDebugMarkerObjectTagInfoEXT and marshal a DebugMarkerObjectTagInfoEXT into it. The VkDebugMarkerObjectTagInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructDebugMarkerObjectTagInfoEXT :: VkDebugMarkerObjectTagInfoEXT -> IO DebugMarkerObjectTagInfoEXT

A function to read a VkDebugMarkerObjectTagInfoEXT and all additional structures in the pointer chain into a DebugMarkerObjectTagInfoEXT.

data DebugMarkerObjectTagInfoEXT

VkDebugMarkerObjectTagInfoEXT - Specify parameters of a tag to attach to an object

Description

The tagName parameter gives a name or identifier to the type of data being tagged. This can be used by debugging layers to easily filter for only data that can be used by that implementation.

Valid Usage (Implicit)

See Also

VkDebugReportObjectTypeEXT, VkStructureType, vkDebugMarkerSetObjectTagEXT

Constructors

DebugMarkerObjectTagInfoEXT

Fields

next :: Maybe SomeVkStruct

objectType :: DebugReportObjectTypeEXT

object :: Word64

tagName :: Word64

tag :: ByteString

cmdDebugMarkerBeginEXT :: CommandBuffer -> DebugMarkerMarkerInfoEXT -> IO ()

vkCmdDebugMarkerBeginEXT - Open a command buffer marker region

Parameters

commandBuffer is the command buffer into which the command is recorded.

pMarkerInfo is a pointer to an instance of the VkDebugMarkerMarkerInfoEXT structure specifying the parameters of the marker region to open.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pMarkerInfo must be a valid pointer to a valid VkDebugMarkerMarkerInfoEXT structure

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkDebugMarkerMarkerInfoEXT

cmdDebugMarkerEndEXT :: CommandBuffer -> IO ()

vkCmdDebugMarkerEndEXT - Close a command buffer marker region

Parameters

commandBuffer is the command buffer into which the command is recorded.

Description

An application may open a marker region in one command buffer and close it in another, or otherwise split marker regions across multiple command buffers or multiple queue submissions. When viewed from the linear series of submissions to a single queue, the calls to vkCmdDebugMarkerBeginEXT and vkCmdDebugMarkerEndEXT must be matched and balanced.

Valid Usage

There must be an outstanding vkCmdDebugMarkerBeginEXT command prior to the vkCmdDebugMarkerEndEXT on the queue that commandBuffer is submitted to

If commandBuffer is a secondary command buffer, there must be an outstanding vkCmdDebugMarkerBeginEXT command recorded to commandBuffer that has not previously been ended by a call to vkCmdDebugMarkerEndEXT.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer

cmdDebugMarkerInsertEXT :: CommandBuffer -> DebugMarkerMarkerInfoEXT -> IO ()

vkCmdDebugMarkerInsertEXT - Insert a marker label into a command buffer

Parameters

commandBuffer is the command buffer into which the command is recorded.

pMarkerInfo is a pointer to an instance of the VkDebugMarkerMarkerInfoEXT structure specifying the parameters of the marker to insert.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pMarkerInfo must be a valid pointer to a valid VkDebugMarkerMarkerInfoEXT structure

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Both            | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCommandBuffer, VkDebugMarkerMarkerInfoEXT

debugMarkerSetObjectNameEXT :: Device -> DebugMarkerObjectNameInfoEXT -> IO ()

vkDebugMarkerSetObjectNameEXT - Give a user-friendly name to an object

Parameters

device is the device that created the object.

pNameInfo is a pointer to an instance of the VkDebugMarkerObjectNameInfoEXT structure specifying the parameters of the name to set on the object.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pNameInfo must be a valid pointer to a valid VkDebugMarkerObjectNameInfoEXT structure

Host Synchronization

Host access to pNameInfo.object must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDebugMarkerObjectNameInfoEXT, VkDevice

debugMarkerSetObjectTagEXT :: Device -> DebugMarkerObjectTagInfoEXT -> IO ()

vkDebugMarkerSetObjectTagEXT - Attach arbitrary data to an object

Parameters

device is the device that created the object.

pTagInfo is a pointer to an instance of the VkDebugMarkerObjectTagInfoEXT structure specifying the parameters of the tag to attach to the object.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pTagInfo must be a valid pointer to a valid VkDebugMarkerObjectTagInfoEXT structure

Host Synchronization

Host access to pTagInfo.object must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDebugMarkerObjectTagInfoEXT, VkDevice

pattern EXT_DEBUG_MARKER_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_DEBUG_MARKER_SPEC_VERSION :: Integral a => a

type DebugReportObjectTypeEXT = VkDebugReportObjectTypeEXT

VkDebugReportObjectTypeEXT - Specify the type of an object handle

Description

'

+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | Vulkan Handle Type        |
| ug_report.VkDebugReportObjectTypeEXT'    |                           |
+==========================================+===========================+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | Unknown\/Undefined Handle |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_UN |                           |
| KNOWN_EXT'                               |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_IN | .DeviceInitialization.VkI |
| STANCE_EXT'                              | nstance'                  |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_PH | .DeviceInitialization.VkP |
| YSICAL_DEVICE_EXT'                       | hysicalDevice'            |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_DE | .DeviceInitialization.VkD |
| VICE_EXT'                                | evice'                    |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_QU | .Queue.VkQueue'           |
| EUE_EXT'                                 |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_SE | .Queue.VkSemaphore'       |
| MAPHORE_EXT'                             |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_CO | .Queue.VkCommandBuffer'   |
| MMAND_BUFFER_EXT'                        |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_FE | .Queue.VkFence'           |
| NCE_EXT'                                 |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_DE | .Memory.VkDeviceMemory'   |
| VICE_MEMORY_EXT'                         |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_BU | .MemoryManagement.VkBuffe |
| FFER_EXT'                                | r'                        |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_IM | .MemoryManagement.VkImage |
| AGE_EXT'                                 | '                         |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_EV | .Event.VkEvent'           |
| ENT_EXT'                                 |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_QU | .Query.VkQueryPool'       |
| ERY_POOL_EXT'                            |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_BU | .BufferView.VkBufferView' |
| FFER_VIEW_EXT'                           |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_IM | .ImageView.VkImageView'   |
| AGE_VIEW_EXT'                            |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_SH | .Shader.VkShaderModule'   |
| ADER_MODULE_EXT'                         |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_PI | .PipelineCache.VkPipeline |
| PELINE_CACHE_EXT'                        | Cache'                    |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_PI | .Pipeline.VkPipelineLayou |
| PELINE_LAYOUT_EXT'                       | t'                        |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_RE | .Pipeline.VkRenderPass'   |
| NDER_PASS_EXT'                           |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_PI | .Pipeline.VkPipeline'     |
| PELINE_EXT'                              |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_DE | .PipelineLayout.VkDescrip |
| SCRIPTOR_SET_LAYOUT_EXT'                 | torSetLayout'             |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_SA | .Sampler.VkSampler'       |
| MPLER_EXT'                               |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_DE | .DescriptorSet.VkDescript |
| SCRIPTOR_POOL_EXT'                       | orPool'                   |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_DE | .DescriptorSet.VkDescript |
| SCRIPTOR_SET_EXT'                        | orSet'                    |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_FR | .Pass.VkFramebuffer'      |
| AMEBUFFER_EXT'                           |                           |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Core10 |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_CO | .CommandPool.VkCommandPoo |
| MMAND_POOL_EXT'                          | l'                        |
+------------------------------------------+---------------------------+
| 'Graphics.Vulkan.C.Extensions.VK_EXT_deb | 'Graphics.Vulkan.C.Extens |
| ug_report.VK_DEBUG_REPORT_OBJECT_TYPE_DE | ions.VK_EXT_debug_report. |
| BUG_REPORT_CALLBACK_EXT_EXT'             | VkDebugReportCallbackEXT' |
+------------------------------------------+---------------------------+

VkDebugReportObjectTypeEXT and Vulkan Handle Relationship

Note

The primary expected use of VK_ERROR_VALIDATION_FAILED_EXT is for validation layer testing. It is not expected that an application would see this error code during normal use of the validation layers.

See Also

VkDebugMarkerObjectNameInfoEXT, VkDebugMarkerObjectTagInfoEXT, vkDebugReportMessageEXT

pattern STRUCTURE_TYPE_DEBUG_MARKER_OBJECT_NAME_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEBUG_MARKER_OBJECT_TAG_INFO_EXT :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEBUG_MARKER_MARKER_INFO_EXT :: a ~ StructureType => a