Graphics.Vulkan.Extensions.VK_EXT_discard_rectangles

Documentation

cmdSetDiscardRectangleEXT :: CommandBuffer -> ("firstDiscardRectangle" ::: Word32) -> ("discardRectangles" ::: Vector Rect2D) -> IO ()

vkCmdSetDiscardRectangleEXT - Set discard rectangles dynamically

Parameters

CommandBuffer is the command buffer into which the command will be recorded.

firstDiscardRectangle is the index of the first discard rectangle whose state is updated by the command.

discardRectangleCount is the number of discard rectangles whose state are updated by the command.

pDiscardRectangles is a pointer to an array of Rect2D structures specifying discard rectangles.

Description

The discard rectangle taken from element i of pDiscardRectangles replace the current state for the discard rectangle index firstDiscardRectangle + i, for i in [0, discardRectangleCount).

Valid Usage

The sum of firstDiscardRectangle and discardRectangleCount must be less than or equal to PhysicalDeviceDiscardRectanglePropertiesEXT::maxDiscardRectangles

The x and y member of offset in each Rect2D element of pDiscardRectangles must be greater than or equal to 0

Evaluation of (offset.x + extent.width) in each Rect2D element of pDiscardRectangles must not cause a signed integer addition overflow

Evaluation of (offset.y + extent.height) in each Rect2D element of pDiscardRectangles must not cause a signed integer addition overflow

Valid Usage (Implicit)

CommandBuffer must be a valid CommandBuffer handle

pDiscardRectangles must be a valid pointer to an array of discardRectangleCount Rect2D structures

CommandBuffer must be in the recording state

The CommandPool that CommandBuffer was allocated from must support graphics operations

discardRectangleCount must be greater than 0

Host Synchronization

Host access to CommandBuffer must be externally synchronized

Host access to the CommandPool that CommandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

CommandBuffer, Rect2D

data PhysicalDeviceDiscardRectanglePropertiesEXT

VkPhysicalDeviceDiscardRectanglePropertiesEXT - Structure describing discard rectangle limits that can be supported by an implementation

Members

The members of the PhysicalDeviceDiscardRectanglePropertiesEXT structure describe the following implementation-dependent limits:

Description

If the PhysicalDeviceDiscardRectanglePropertiesEXT structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

StructureType

Constructors

PhysicalDeviceDiscardRectanglePropertiesEXT

Fields

maxDiscardRectangles :: Word32

maxDiscardRectangles is the maximum number of active discard rectangles that can be specified.

data PipelineDiscardRectangleStateCreateInfoEXT

VkPipelineDiscardRectangleStateCreateInfoEXT - Structure specifying discard rectangle

Valid Usage (Implicit)

See Also

DiscardRectangleModeEXT, PipelineDiscardRectangleStateCreateFlagsEXT, Rect2D, StructureType

Constructors

PipelineDiscardRectangleStateCreateInfoEXT

Fields

flags :: PipelineDiscardRectangleStateCreateFlagsEXT

Flags must be 0

discardRectangleMode :: DiscardRectangleModeEXT

discardRectangleMode must be a valid DiscardRectangleModeEXT value

discardRectangles :: Either Word32 (Vector Rect2D)

pDiscardRectangles is a pointer to an array of Rect2D structures, defining the discard rectangles. If the discard rectangle state is dynamic, this member is ignored.

newtype PipelineDiscardRectangleStateCreateFlagsEXT

VkPipelineDiscardRectangleStateCreateFlagsEXT - Reserved for future use

Description

PipelineDiscardRectangleStateCreateFlagsEXT is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

PipelineDiscardRectangleStateCreateInfoEXT

Constructors

PipelineDiscardRectangleStateCreateFlagsEXT Flags

newtype DiscardRectangleModeEXT

VkDiscardRectangleModeEXT - Specify the discard rectangle mode

See Also

PipelineDiscardRectangleStateCreateInfoEXT

Constructors

DiscardRectangleModeEXT Int32

Bundled Patterns

pattern DISCARD_RECTANGLE_MODE_INCLUSIVE_EXT :: DiscardRectangleModeEXT
DISCARD_RECTANGLE_MODE_INCLUSIVE_EXT specifies that a fragment within any discard rectangle satisfies the test.

pattern DISCARD_RECTANGLE_MODE_EXCLUSIVE_EXT :: DiscardRectangleModeEXT
DISCARD_RECTANGLE_MODE_EXCLUSIVE_EXT specifies that a fragment not within any of the discard rectangles satisfies the test.

type EXT_DISCARD_RECTANGLES_SPEC_VERSION = 1

pattern EXT_DISCARD_RECTANGLES_SPEC_VERSION :: forall a. Integral a => a

type EXT_DISCARD_RECTANGLES_EXTENSION_NAME = "VK_EXT_discard_rectangles"

pattern EXT_DISCARD_RECTANGLES_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a