Graphics.Vulkan.Extensions.VK_EXT_display_control

Documentation

displayPowerControlEXT :: Device -> DisplayKHR -> DisplayPowerInfoEXT -> IO ()

vkDisplayPowerControlEXT - Set the power state of a display

Parameters

Device is a logical device associated with Display.

Display is the display whose power state is modified.

pDisplayPowerInfo is a DisplayPowerInfoEXT structure specifying the new power state of Display.

Valid Usage (Implicit)

Device must be a valid Device handle

Display must be a valid DisplayKHR handle

pDisplayPowerInfo must be a valid pointer to a valid DisplayPowerInfoEXT structure

Both of Device, and Display must have been created, allocated, or retrieved from the same PhysicalDevice

Return Codes

Success

SUCCESS

See Also

Device, DisplayKHR, DisplayPowerInfoEXT

registerDeviceEventEXT :: Device -> DeviceEventInfoEXT -> ("allocator" ::: Maybe AllocationCallbacks) -> IO Fence

vkRegisterDeviceEventEXT - Signal a fence when a device event occurs

Parameters

Device is a logical device on which the event may occur.

pDeviceEventInfo is a pointer to a DeviceEventInfoEXT structure describing the event of interest to the application.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pFence is a pointer to a handle in which the resulting fence object is returned.

Valid Usage (Implicit)

Device must be a valid Device handle

pDeviceEventInfo must be a valid pointer to a valid DeviceEventInfoEXT structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pFence must be a valid pointer to a Fence handle

Return Codes

Success

SUCCESS

See Also

AllocationCallbacks, Device, DeviceEventInfoEXT, Fence

registerDisplayEventEXT :: Device -> DisplayKHR -> DisplayEventInfoEXT -> ("allocator" ::: Maybe AllocationCallbacks) -> IO Fence

vkRegisterDisplayEventEXT - Signal a fence when a display event occurs

Parameters

Device is a logical device associated with Display

Display is the display on which the event may occur.

pDisplayEventInfo is a pointer to a DisplayEventInfoEXT structure describing the event of interest to the application.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pFence is a pointer to a handle in which the resulting fence object is returned.

Valid Usage (Implicit)

Device must be a valid Device handle

Display must be a valid DisplayKHR handle

pDisplayEventInfo must be a valid pointer to a valid DisplayEventInfoEXT structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pFence must be a valid pointer to a Fence handle

Both of Device, and Display must have been created, allocated, or retrieved from the same PhysicalDevice

Return Codes

Success

SUCCESS

See Also

AllocationCallbacks, Device, DisplayEventInfoEXT, DisplayKHR, Fence

getSwapchainCounterEXT :: Device -> SwapchainKHR -> SurfaceCounterFlagBitsEXT -> IO ("counterValue" ::: Word64)

vkGetSwapchainCounterEXT - Query the current value of a surface counter

Parameters

Device is the Device associated with swapchain.

swapchain is the swapchain from which to query the counter value.

counter is the counter to query.

pCounterValue will return the current value of the counter.

Description

If a counter is not available because the swapchain is out of date, the implementation may return ERROR_OUT_OF_DATE_KHR.

Valid Usage

One or more present commands on swapchain must have been processed by the presentation engine.

Valid Usage (Implicit)

Device must be a valid Device handle

swapchain must be a valid SwapchainKHR handle

counter must be a valid SurfaceCounterFlagBitsEXT value

pCounterValue must be a valid pointer to a uint64_t value

Both of Device, and swapchain must have been created, allocated, or retrieved from the same Instance

Return Codes

Success

SUCCESS

Failure

ERROR_DEVICE_LOST

ERROR_OUT_OF_DATE_KHR

See Also

Device, SurfaceCounterFlagBitsEXT, SwapchainKHR

data DisplayPowerInfoEXT

VkDisplayPowerInfoEXT - Describe the power state of a display

Valid Usage (Implicit)

See Also

DisplayPowerStateEXT, StructureType, displayPowerControlEXT

Constructors

DisplayPowerInfoEXT

Fields

powerState :: DisplayPowerStateEXT

powerState must be a valid DisplayPowerStateEXT value

data DeviceEventInfoEXT

VkDeviceEventInfoEXT - Describe a device event to create

Valid Usage (Implicit)

See Also

DeviceEventTypeEXT, StructureType, registerDeviceEventEXT

Constructors

DeviceEventInfoEXT

Fields

deviceEvent :: DeviceEventTypeEXT

deviceEvent must be a valid DeviceEventTypeEXT value

data DisplayEventInfoEXT

VkDisplayEventInfoEXT - Describe a display event to create

Valid Usage (Implicit)

See Also

DisplayEventTypeEXT, StructureType, registerDisplayEventEXT

Constructors

DisplayEventInfoEXT

Fields

displayEvent :: DisplayEventTypeEXT

displayEvent must be a valid DisplayEventTypeEXT value

data SwapchainCounterCreateInfoEXT

VkSwapchainCounterCreateInfoEXT - Specify the surface counters desired

Valid Usage

The bits in surfaceCounters must be supported by SwapchainCreateInfoKHR::surface, as reported by getPhysicalDeviceSurfaceCapabilities2EXT.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_SWAPCHAIN_COUNTER_CREATE_INFO_EXT

surfaceCounters must be a valid combination of SurfaceCounterFlagBitsEXT values

See Also

StructureType, SurfaceCounterFlagsEXT

Constructors

SwapchainCounterCreateInfoEXT

Fields

surfaceCounters :: SurfaceCounterFlagsEXT

surfaceCounters is a bitmask of SurfaceCounterFlagBitsEXT specifying surface counters to enable for the swapchain.

newtype DisplayPowerStateEXT

VkDisplayPowerStateEXT - Possible power states for a display

See Also

DisplayPowerInfoEXT

Constructors

DisplayPowerStateEXT Int32

Bundled Patterns

pattern DISPLAY_POWER_STATE_OFF_EXT :: DisplayPowerStateEXT
DISPLAY_POWER_STATE_OFF_EXT specifies that the display is powered down.

pattern DISPLAY_POWER_STATE_SUSPEND_EXT :: DisplayPowerStateEXT
DISPLAY_POWER_STATE_SUSPEND_EXT specifies that the display is put into a low power mode, from which it may be able to transition back to DISPLAY_POWER_STATE_ON_EXT more quickly than if it were in DISPLAY_POWER_STATE_OFF_EXT. This state may be the same as DISPLAY_POWER_STATE_OFF_EXT.

pattern DISPLAY_POWER_STATE_ON_EXT :: DisplayPowerStateEXT
DISPLAY_POWER_STATE_ON_EXT specifies that the display is powered on.

newtype DeviceEventTypeEXT

VkDeviceEventTypeEXT - Events that can occur on a device object

See Also

DeviceEventInfoEXT

Constructors

DeviceEventTypeEXT Int32

Bundled Patterns

pattern DEVICE_EVENT_TYPE_DISPLAY_HOTPLUG_EXT :: DeviceEventTypeEXT
DEVICE_EVENT_TYPE_DISPLAY_HOTPLUG_EXT specifies that the fence is signaled when a display is plugged into or unplugged from the specified device. Applications can use this notification to determine when they need to re-enumerate the available displays on a device.

newtype DisplayEventTypeEXT

VkDisplayEventTypeEXT - Events that can occur on a display object

See Also

DisplayEventInfoEXT

Constructors

DisplayEventTypeEXT Int32

Bundled Patterns

pattern DISPLAY_EVENT_TYPE_FIRST_PIXEL_OUT_EXT :: DisplayEventTypeEXT
DISPLAY_EVENT_TYPE_FIRST_PIXEL_OUT_EXT specifies that the fence is signaled when the first pixel of the next display refresh cycle leaves the display engine for the display.

type EXT_DISPLAY_CONTROL_SPEC_VERSION = 1

pattern EXT_DISPLAY_CONTROL_SPEC_VERSION :: forall a. Integral a => a

type EXT_DISPLAY_CONTROL_EXTENSION_NAME = "VK_EXT_display_control"

pattern EXT_DISPLAY_CONTROL_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype DisplayKHR

VkDisplayKHR - Opaque handle to a display object

See Also

DisplayPlanePropertiesKHR, DisplayPropertiesKHR, acquireXlibDisplayEXT, createDisplayModeKHR, displayPowerControlEXT, getDisplayModeProperties2KHR, getDisplayModePropertiesKHR, getDisplayPlaneSupportedDisplaysKHR, getRandROutputDisplayEXT, registerDisplayEventEXT, releaseDisplayEXT

Constructors

DisplayKHR Word64

newtype SwapchainKHR

VkSwapchainKHR - Opaque handle to a swapchain object

Description

A swapchain is an abstraction for an array of presentable images that are associated with a surface. The presentable images are represented by Image objects created by the platform. One image (which can be an array image for multiview/stereoscopic-3D surfaces) is displayed at a time, but multiple images can be queued for presentation. An application renders to the image, and then queues the image for presentation to the surface.

A native window cannot be associated with more than one non-retired swapchain at a time. Further, swapchains cannot be created for native windows that have a non-Vulkan graphics API surface associated with them.

Note

The presentation engine is an abstraction for the platform’s compositor or display engine.

The presentation engine may be synchronous or asynchronous with respect to the application and/or logical device.

Some implementations may use the device’s graphics queue or dedicated presentation hardware to perform presentation.

The presentable images of a swapchain are owned by the presentation engine. An application can acquire use of a presentable image from the presentation engine. Use of a presentable image must occur only after the image is returned by acquireNextImageKHR, and before it is presented by queuePresentKHR. This includes transitioning the image layout and rendering commands.

An application can acquire use of a presentable image with acquireNextImageKHR. After acquiring a presentable image and before modifying it, the application must use a synchronization primitive to ensure that the presentation engine has finished reading from the image. The application can then transition the image’s layout, queue rendering commands to it, etc. Finally, the application presents the image with queuePresentKHR, which releases the acquisition of the image.

The presentation engine controls the order in which presentable images are acquired for use by the application.

Note

This allows the platform to handle situations which require out-of-order return of images after presentation. At the same time, it allows the application to generate command buffers referencing all of the images in the swapchain at initialization time, rather than in its main loop.

See Also

AcquireNextImageInfoKHR, BindImageMemorySwapchainInfoKHR, ImageSwapchainCreateInfoKHR, PresentInfoKHR, SwapchainCreateInfoKHR, acquireFullScreenExclusiveModeEXT, acquireNextImageKHR, createSharedSwapchainsKHR, createSwapchainKHR, destroySwapchainKHR, getPastPresentationTimingGOOGLE, getRefreshCycleDurationGOOGLE, getSwapchainCounterEXT, getSwapchainImagesKHR, getSwapchainStatusKHR, releaseFullScreenExclusiveModeEXT, setHdrMetadataEXT, setLocalDimmingAMD

Constructors

SwapchainKHR Word64

newtype SurfaceCounterFlagBitsEXT

VkSurfaceCounterFlagBitsEXT - Surface-relative counter types

See Also

SurfaceCounterFlagsEXT, getSwapchainCounterEXT

Constructors

SurfaceCounterFlagBitsEXT Flags

Bundled Patterns

pattern SURFACE_COUNTER_VBLANK_EXT :: SurfaceCounterFlagBitsEXT
SURFACE_COUNTER_VBLANK_EXT specifies a counter incrementing once every time a vertical blanking period occurs on the display associated with the surface.

type SurfaceCounterFlagsEXT = SurfaceCounterFlagBitsEXT