Graphics.Vulkan.Extensions.VK_EXT_headless_surface

Documentation

createHeadlessSurfaceEXT :: forall io. MonadIO io => Instance -> HeadlessSurfaceCreateInfoEXT -> ("allocator" ::: Maybe AllocationCallbacks) -> io SurfaceKHR

vkCreateHeadlessSurfaceEXT - Create a headless SurfaceKHR object

Parameters

instance is the instance to associate the surface with.

pCreateInfo is a pointer to a HeadlessSurfaceCreateInfoEXT structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface is a pointer to a SurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

instance must be a valid Instance handle

pCreateInfo must be a valid pointer to a valid HeadlessSurfaceCreateInfoEXT structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pSurface must be a valid pointer to a SurfaceKHR handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, HeadlessSurfaceCreateInfoEXT, Instance, SurfaceKHR

data HeadlessSurfaceCreateInfoEXT

VkHeadlessSurfaceCreateInfoEXT - Structure specifying parameters of a newly created headless surface object

Valid Usage (Implicit)

See Also

HeadlessSurfaceCreateFlagsEXT, StructureType, createHeadlessSurfaceEXT

Constructors

HeadlessSurfaceCreateInfoEXT

Fields

flags :: HeadlessSurfaceCreateFlagsEXT

flags must be 0

newtype HeadlessSurfaceCreateFlagsEXT

Constructors

HeadlessSurfaceCreateFlagsEXT Flags

type EXT_HEADLESS_SURFACE_SPEC_VERSION = 1

pattern EXT_HEADLESS_SURFACE_SPEC_VERSION :: forall a. Integral a => a

type EXT_HEADLESS_SURFACE_EXTENSION_NAME = "VK_EXT_headless_surface"

pattern EXT_HEADLESS_SURFACE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype SurfaceKHR

VkSurfaceKHR - Opaque handle to a surface object

Description

The VK_KHR_surface extension declares the SurfaceKHR object, and provides a function for destroying SurfaceKHR objects. Separate platform-specific extensions each provide a function for creating a SurfaceKHR object for the respective platform. From the application’s perspective this is an opaque handle, just like the handles of other Vulkan objects.

See Also

PhysicalDeviceSurfaceInfo2KHR, SwapchainCreateInfoKHR, createAndroidSurfaceKHR, createDisplayPlaneSurfaceKHR, createHeadlessSurfaceEXT, createIOSSurfaceMVK, createImagePipeSurfaceFUCHSIA, createMacOSSurfaceMVK, createMetalSurfaceEXT, createStreamDescriptorSurfaceGGP, createViSurfaceNN, createWaylandSurfaceKHR, createWin32SurfaceKHR, createXcbSurfaceKHR, createXlibSurfaceKHR, destroySurfaceKHR, getDeviceGroupSurfacePresentModesKHR, getPhysicalDevicePresentRectanglesKHR, getPhysicalDeviceSurfaceCapabilities2EXT, getPhysicalDeviceSurfaceCapabilitiesKHR, getPhysicalDeviceSurfaceFormatsKHR, getPhysicalDeviceSurfacePresentModesKHR, getPhysicalDeviceSurfaceSupportKHR

Constructors

SurfaceKHR Word64