Graphics.Vulkan.Extensions.VK_EXT_memory_budget

Documentation

data PhysicalDeviceMemoryBudgetPropertiesEXT

VkPhysicalDeviceMemoryBudgetPropertiesEXT - Structure specifying physical device memory budget and usage

Description

The values returned in this structure are not invariant. The heapBudget and heapUsage values must be zero for array elements greater than or equal to PhysicalDeviceMemoryProperties::memoryHeapCount. The heapBudget value must be non-zero for array elements less than PhysicalDeviceMemoryProperties::memoryHeapCount. The heapBudget value must be less than or equal to MemoryHeap::size for each heap.

Valid Usage (Implicit)

See Also

DeviceSize, StructureType

Constructors

PhysicalDeviceMemoryBudgetPropertiesEXT

Fields

heapBudget :: Vector DeviceSize

heapBudget is an array of MAX_MEMORY_HEAPS DeviceSize values in which memory budgets are returned, with one element for each memory heap. A heap’s budget is a rough estimate of how much memory the process can allocate from that heap before allocations may fail or cause performance degradation. The budget includes any currently allocated device memory.

heapUsage :: Vector DeviceSize

heapUsage is an array of MAX_MEMORY_HEAPS DeviceSize values in which memory usages are returned, with one element for each memory heap. A heap’s usage is an estimate of how much memory the process is currently using in that heap.

type EXT_MEMORY_BUDGET_SPEC_VERSION = 1

pattern EXT_MEMORY_BUDGET_SPEC_VERSION :: forall a. Integral a => a

type EXT_MEMORY_BUDGET_EXTENSION_NAME = "VK_EXT_memory_budget"

pattern EXT_MEMORY_BUDGET_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a