Graphics.Vulkan.Extensions.VK_EXT_metal_surface

Documentation

type MetalSurfaceCreateFlagsEXT = VkMetalSurfaceCreateFlagsEXT

withCStructMetalSurfaceCreateInfoEXT :: MetalSurfaceCreateInfoEXT -> (VkMetalSurfaceCreateInfoEXT -> IO a) -> IO a

A function to temporarily allocate memory for a VkMetalSurfaceCreateInfoEXT and marshal a MetalSurfaceCreateInfoEXT into it. The VkMetalSurfaceCreateInfoEXT is only valid inside the provided computation and must not be returned out of it.

fromCStructMetalSurfaceCreateInfoEXT :: VkMetalSurfaceCreateInfoEXT -> IO MetalSurfaceCreateInfoEXT

A function to read a VkMetalSurfaceCreateInfoEXT and all additional structures in the pointer chain into a MetalSurfaceCreateInfoEXT.

data MetalSurfaceCreateInfoEXT

VkMetalSurfaceCreateInfoEXT - Structure specifying parameters of a newly created Metal surface object

Valid Usage (Implicit)

See Also

VkMetalSurfaceCreateFlagsEXT, VkStructureType, vkCreateMetalSurfaceEXT

Constructors

MetalSurfaceCreateInfoEXT

Fields

next :: Maybe SomeVkStruct

flags :: MetalSurfaceCreateFlagsEXT

layer :: Ptr CAMetalLayer

createMetalSurfaceEXT :: Instance -> MetalSurfaceCreateInfoEXT -> Maybe AllocationCallbacks -> IO SurfaceKHR

vkCreateMetalSurfaceEXT - Create a VkSurfaceKHR object for CAMetalLayer

Parameters

instance is the instance with which to associate the surface.

pCreateInfo is a pointer to an instance of the VkMetalSurfaceCreateInfoEXT structure containing the parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface points to a VkSurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

instance must be a valid VkInstance handle

pCreateInfo must be a valid pointer to a valid VkMetalSurfaceCreateInfoEXT structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pSurface must be a valid pointer to a VkSurfaceKHR handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_NATIVE_WINDOW_IN_USE_KHR

See Also

VkAllocationCallbacks, VkInstance, VkMetalSurfaceCreateInfoEXT, VkSurfaceKHR

pattern EXT_METAL_SURFACE_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern EXT_METAL_SURFACE_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_METAL_SURFACE_CREATE_INFO_EXT :: a ~ StructureType => a