Graphics.Vulkan.Extensions.VK_EXT_pipeline_creation_feedback

Documentation

data PipelineCreationFeedbackEXT

VkPipelineCreationFeedbackEXT - Feedback about the creation of a pipeline or pipeline stage

Description

If the PIPELINE_CREATION_FEEDBACK_VALID_BIT_EXT is not set in Flags, an implementation must not set any other bits in Flags, and all other PipelineCreationFeedbackEXT data members are undefined.

See Also

PipelineCreationFeedbackCreateInfoEXT, PipelineCreationFeedbackFlagBitsEXT, PipelineCreationFeedbackFlagsEXT

Constructors

PipelineCreationFeedbackEXT

Fields

flags :: PipelineCreationFeedbackFlagsEXT

Flags is a bitmask of PipelineCreationFeedbackFlagBitsEXT providing feedback about the creation of a pipeline or of a pipeline stage.

duration :: Word64

duration is the duration spent creating a pipeline or pipeline stage in nanoseconds.

data PipelineCreationFeedbackCreateInfoEXT

VkPipelineCreationFeedbackCreateInfoEXT - Request for feedback about the creation of a pipeline

Description

An implementation should write pipeline creation feedback to pPipelineCreationFeedback and may write pipeline stage creation feedback to pPipelineStageCreationFeedbacks. An implementation must set or clear the PIPELINE_CREATION_FEEDBACK_VALID_BIT_EXT in 'PipelineCreationFeedbackEXT'::'Graphics.Vulkan.Core10.BaseType.Flags' for pPipelineCreationFeedback and every element of pPipelineStageCreationFeedbacks.

Note

One common scenario for an implementation to skip per-stage feedback is when PIPELINE_CREATION_FEEDBACK_APPLICATION_PIPELINE_CACHE_HIT_BIT_EXT is set in pPipelineCreationFeedback.

When chained to RayTracingPipelineCreateInfoNV or GraphicsPipelineCreateInfo, the i element of pPipelineStageCreationFeedbacks corresponds to the i element of RayTracingPipelineCreateInfoNV::pStages or GraphicsPipelineCreateInfo::pStages. When chained to ComputePipelineCreateInfo, the first element of pPipelineStageCreationFeedbacks corresponds to ComputePipelineCreateInfo::stage.

Valid Usage

When chained to GraphicsPipelineCreateInfo, PipelineCreationFeedbackEXT::pipelineStageCreationFeedbackCount must equal GraphicsPipelineCreateInfo::stageCount

When chained to ComputePipelineCreateInfo, PipelineCreationFeedbackEXT::pipelineStageCreationFeedbackCount must equal 1

When chained to RayTracingPipelineCreateInfoNV, PipelineCreationFeedbackEXT::pipelineStageCreationFeedbackCount must equal RayTracingPipelineCreateInfoNV::stageCount

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PIPELINE_CREATION_FEEDBACK_CREATE_INFO_EXT

pPipelineCreationFeedback must be a valid pointer to a PipelineCreationFeedbackEXT structure

pPipelineStageCreationFeedbacks must be a valid pointer to an array of pipelineStageCreationFeedbackCount PipelineCreationFeedbackEXT structures

pipelineStageCreationFeedbackCount must be greater than 0

See Also

ComputePipelineCreateInfo, GraphicsPipelineCreateInfo, PipelineCreationFeedbackEXT, RayTracingPipelineCreateInfoNV, StructureType

Constructors

PipelineCreationFeedbackCreateInfoEXT

Fields

pipelineCreationFeedback :: Ptr PipelineCreationFeedbackEXT

pPipelineCreationFeedback is a pointer to a PipelineCreationFeedbackEXT structure.

pipelineStageCreationFeedbackCount :: Word32

pipelineStageCreationFeedbackCount is the number of elements in pPipelineStageCreationFeedbacks.

pipelineStageCreationFeedbacks :: Ptr PipelineCreationFeedbackEXT

pPipelineStageCreationFeedbacks is a pointer to an array of pipelineStageCreationFeedbackCount PipelineCreationFeedbackEXT structures.

newtype PipelineCreationFeedbackFlagBitsEXT

VkPipelineCreationFeedbackFlagBitsEXT - Bitmask specifying pipeline or pipeline stage creation feedback

See Also

PipelineCreationFeedbackCreateInfoEXT, PipelineCreationFeedbackEXT, PipelineCreationFeedbackFlagsEXT

Constructors

PipelineCreationFeedbackFlagBitsEXT Flags

Bundled Patterns

pattern PIPELINE_CREATION_FEEDBACK_VALID_BIT_EXT :: PipelineCreationFeedbackFlagBitsEXT
PIPELINE_CREATION_FEEDBACK_VALID_BIT_EXT indicates that the feedback information is valid.

pattern PIPELINE_CREATION_FEEDBACK_APPLICATION_PIPELINE_CACHE_HIT_BIT_EXT :: PipelineCreationFeedbackFlagBitsEXT
PIPELINE_CREATION_FEEDBACK_APPLICATION_PIPELINE_CACHE_HIT_BIT_EXT indicates that a readily usable pipeline or pipeline stage was found in the PipelineCache specified by the application in the pipeline creation command.

An implementation should set the PIPELINE_CREATION_FEEDBACK_APPLICATION_PIPELINE_CACHE_HIT_BIT_EXT bit if it was able to avoid the large majority of pipeline or pipeline stage creation work by using the PipelineCache parameter of createGraphicsPipelines, createRayTracingPipelinesNV, or createComputePipelines. When an implementation sets this bit for the entire pipeline, it may leave it unset for any stage.

Note

Implementations are encouraged to provide a meaningful signal to applications using this bit. The intention is to communicate to the application that the pipeline or pipeline stage was created "as fast as it gets" using the pipeline cache provided by the application. If an implementation uses an internal cache, it is discouraged from setting this bit as the feedback would be unactionable.

pattern PIPELINE_CREATION_FEEDBACK_BASE_PIPELINE_ACCELERATION_BIT_EXT :: PipelineCreationFeedbackFlagBitsEXT
PIPELINE_CREATION_FEEDBACK_BASE_PIPELINE_ACCELERATION_BIT_EXT indicates that the base pipeline specified by the basePipelineHandle or basePipelineIndex member of the Vk*PipelineCreateInfo structure was used to accelerate the creation of the pipeline.

An implementation should set the PIPELINE_CREATION_FEEDBACK_BASE_PIPELINE_ACCELERATION_BIT_EXT bit if it was able to avoid a significant amount of work by using the base pipeline.

Note

While "significant amount of work" is subjective, implementations are encouraged to provide a meaningful signal to applications using this bit. For example, a 1% reduction in duration may not warrant setting this bit, while a 50% reduction would.

type PipelineCreationFeedbackFlagsEXT = PipelineCreationFeedbackFlagBitsEXT

type EXT_PIPELINE_CREATION_FEEDBACK_SPEC_VERSION = 1

pattern EXT_PIPELINE_CREATION_FEEDBACK_SPEC_VERSION :: forall a. Integral a => a

type EXT_PIPELINE_CREATION_FEEDBACK_EXTENSION_NAME = "VK_EXT_pipeline_creation_feedback"

pattern EXT_PIPELINE_CREATION_FEEDBACK_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a