Graphics.Vulkan.Extensions.VK_EXT_vertex_attribute_divisor

Documentation

data VertexInputBindingDivisorDescriptionEXT

VkVertexInputBindingDivisorDescriptionEXT - Structure specifying a divisor used in instanced rendering

Description

If this structure is not used to define a divisor value for an attribute then the divisor has a logical default value of 1.

Valid Usage

binding must be less than PhysicalDeviceLimits::maxVertexInputBindings

If the vertexAttributeInstanceRateZeroDivisor feature is not enabled, divisor must not be 0

If the vertexAttributeInstanceRateDivisor feature is not enabled, divisor must be 1

divisor must be a value between 0 and PhysicalDeviceVertexAttributeDivisorPropertiesEXT::maxVertexAttribDivisor, inclusive.

VertexInputBindingDescription::inputRate must be of type VERTEX_INPUT_RATE_INSTANCE for this binding.

See Also

PipelineVertexInputDivisorStateCreateInfoEXT

Constructors

VertexInputBindingDivisorDescriptionEXT

Fields

binding :: Word32

binding is the binding number for which the divisor is specified.

divisor :: Word32

divisor is the number of successive instances that will use the same value of the vertex attribute when instanced rendering is enabled. For example, if the divisor is N, the same vertex attribute will be applied to N successive instances before moving on to the next vertex attribute. The maximum value of divisor is implementation dependent and can be queried using PhysicalDeviceVertexAttributeDivisorPropertiesEXT::maxVertexAttribDivisor. A value of 0 can be used for the divisor if the vertexAttributeInstanceRateZeroDivisor feature is enabled. In this case, the same vertex attribute will be applied to all instances.

data PipelineVertexInputDivisorStateCreateInfoEXT

VkPipelineVertexInputDivisorStateCreateInfoEXT - Structure specifying vertex attributes assignment during instanced rendering

Valid Usage (Implicit)

See Also

StructureType, VertexInputBindingDivisorDescriptionEXT

Constructors

PipelineVertexInputDivisorStateCreateInfoEXT

Fields

vertexBindingDivisors :: Vector VertexInputBindingDivisorDescriptionEXT

pVertexBindingDivisors must be a valid pointer to an array of vertexBindingDivisorCount VertexInputBindingDivisorDescriptionEXT structures

data PhysicalDeviceVertexAttributeDivisorPropertiesEXT

VkPhysicalDeviceVertexAttributeDivisorPropertiesEXT - Structure describing max value of vertex attribute divisor that can be supported by an implementation

Members

The members of the PhysicalDeviceVertexAttributeDivisorPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the PhysicalDeviceVertexAttributeDivisorPropertiesEXT structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

StructureType

Constructors

PhysicalDeviceVertexAttributeDivisorPropertiesEXT

Fields

maxVertexAttribDivisor :: Word32

maxVertexAttribDivisor is the maximum value of the number of instances that will repeat the value of vertex attribute data when instanced rendering is enabled.

data PhysicalDeviceVertexAttributeDivisorFeaturesEXT

VkPhysicalDeviceVertexAttributeDivisorFeaturesEXT - Structure describing if fetching of vertex attribute may be repeated for instanced rendering

Description

If the PhysicalDeviceVertexAttributeDivisorFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating the implementation-dependent behavior. PhysicalDeviceVertexAttributeDivisorFeaturesEXT can also be included in pNext chain of DeviceCreateInfo to enable the feature.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceVertexAttributeDivisorFeaturesEXT

Fields

vertexAttributeInstanceRateDivisor :: Bool

vertexAttributeInstanceRateDivisor specifies whether vertex attribute fetching may be repeated in case of instanced rendering.

vertexAttributeInstanceRateZeroDivisor :: Bool

vertexAttributeInstanceRateZeroDivisor specifies whether a zero value for VertexInputBindingDivisorDescriptionEXT::divisor is supported.

type EXT_VERTEX_ATTRIBUTE_DIVISOR_SPEC_VERSION = 3

pattern EXT_VERTEX_ATTRIBUTE_DIVISOR_SPEC_VERSION :: forall a. Integral a => a

type EXT_VERTEX_ATTRIBUTE_DIVISOR_EXTENSION_NAME = "VK_EXT_vertex_attribute_divisor"

pattern EXT_VERTEX_ATTRIBUTE_DIVISOR_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a