Graphics.Vulkan.Extensions.VK_KHR_external_fence_win32

Documentation

withCStructExportFenceWin32HandleInfoKHR :: ExportFenceWin32HandleInfoKHR -> (VkExportFenceWin32HandleInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkExportFenceWin32HandleInfoKHR and marshal a ExportFenceWin32HandleInfoKHR into it. The VkExportFenceWin32HandleInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructExportFenceWin32HandleInfoKHR :: VkExportFenceWin32HandleInfoKHR -> IO ExportFenceWin32HandleInfoKHR

A function to read a VkExportFenceWin32HandleInfoKHR and all additional structures in the pointer chain into a ExportFenceWin32HandleInfoKHR.

data ExportFenceWin32HandleInfoKHR

VkExportFenceWin32HandleInfoKHR - Structure specifying additional attributes of Windows handles exported from a fence

Description

If this structure is not present, or if pAttributes is set to NULL, default security descriptor values will be used, and child processes created by the application will not inherit the handle, as described in the MSDN documentation for “Synchronization Object Security and Access Rights”1. Further, if the structure is not present, the access rights will be

DXGI_SHARED_RESOURCE_READ | DXGI_SHARED_RESOURCE_WRITE

for handles of the following types:

VK_EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_BIT

1

https://msdn.microsoft.com/en-us/library/windows/desktop/ms686670.aspx

Valid Usage

If VkExportFenceCreateInfo::handleTypes does not include VK_EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_BIT, VkExportFenceWin32HandleInfoKHR must not be in the pNext chain of VkFenceCreateInfo.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_EXPORT_FENCE_WIN32_HANDLE_INFO_KHR

If pAttributes is not NULL, pAttributes must be a valid pointer to a valid SECURITY_ATTRIBUTES value

See Also

VkStructureType

Constructors

ExportFenceWin32HandleInfoKHR

Fields

next :: Maybe SomeVkStruct

attributes :: Ptr SECURITY_ATTRIBUTES

dwAccess :: DWORD

name :: LPCWSTR

withCStructFenceGetWin32HandleInfoKHR :: FenceGetWin32HandleInfoKHR -> (VkFenceGetWin32HandleInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkFenceGetWin32HandleInfoKHR and marshal a FenceGetWin32HandleInfoKHR into it. The VkFenceGetWin32HandleInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructFenceGetWin32HandleInfoKHR :: VkFenceGetWin32HandleInfoKHR -> IO FenceGetWin32HandleInfoKHR

A function to read a VkFenceGetWin32HandleInfoKHR and all additional structures in the pointer chain into a FenceGetWin32HandleInfoKHR.

data FenceGetWin32HandleInfoKHR

VkFenceGetWin32HandleInfoKHR - Structure describing a Win32 handle fence export operation

Description

The properties of the handle returned depend on the value of handleType. See VkExternalFenceHandleTypeFlagBits for a description of the properties of the defined external fence handle types.

Valid Usage

handleType must have been included in VkExportFenceCreateInfo::handleTypes when the fence’s current payload was created.

If handleType is defined as an NT handle, vkGetFenceWin32HandleKHR must be called no more than once for each valid unique combination of fence and handleType.

fence must not currently have its payload replaced by an imported payload as described below in Importing Fence Payloads unless that imported payload’s handle type was included in VkExternalFenceProperties::exportFromImportedHandleTypes for handleType.

If handleType refers to a handle type with copy payload transference semantics, fence must be signaled, or have an associated fence signal operation pending execution.

handleType must be defined as an NT handle or a global share handle.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_FENCE_GET_WIN32_HANDLE_INFO_KHR

pNext must be NULL

fence must be a valid VkFence handle

handleType must be a valid VkExternalFenceHandleTypeFlagBits value

See Also

VkExternalFenceHandleTypeFlagBits, VkFence, VkStructureType, vkGetFenceWin32HandleKHR

Constructors

FenceGetWin32HandleInfoKHR

Fields

next :: Maybe SomeVkStruct

fence :: Fence

handleType :: ExternalFenceHandleTypeFlagBits

withCStructImportFenceWin32HandleInfoKHR :: ImportFenceWin32HandleInfoKHR -> (VkImportFenceWin32HandleInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkImportFenceWin32HandleInfoKHR and marshal a ImportFenceWin32HandleInfoKHR into it. The VkImportFenceWin32HandleInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructImportFenceWin32HandleInfoKHR :: VkImportFenceWin32HandleInfoKHR -> IO ImportFenceWin32HandleInfoKHR

A function to read a VkImportFenceWin32HandleInfoKHR and all additional structures in the pointer chain into a ImportFenceWin32HandleInfoKHR.

data ImportFenceWin32HandleInfoKHR

VkImportFenceWin32HandleInfoKHR - (None)

Description

The handle types supported by handleType are:

+-----------------------+-----------------------+-----------------------+
| Handle Type           | Transference          | Permanence Supported  |
+=======================+=======================+=======================+
| 'Graphics.Vulkan.C.Co | Reference             | Temporary,Permanent   |
| re11.Promoted_from_VK |                       |                       |
| _KHR_external_fence_c |                       |                       |
| apabilities.VK_EXTERN |                       |                       |
| AL_FENCE_HANDLE_TYPE_ |                       |                       |
| OPAQUE_WIN32_BIT'     |                       |                       |
+-----------------------+-----------------------+-----------------------+
| 'Graphics.Vulkan.C.Co | Reference             | Temporary,Permanent   |
| re11.Promoted_from_VK |                       |                       |
| _KHR_external_fence_c |                       |                       |
| apabilities.VK_EXTERN |                       |                       |
| AL_FENCE_HANDLE_TYPE_ |                       |                       |
| OPAQUE_WIN32_KMT_BIT' |                       |                       |
+-----------------------+-----------------------+-----------------------+

Handle Types Supported by
'Graphics.Vulkan.C.Extensions.VK_KHR_external_fence_win32.VkImportFenceWin32HandleInfoKHR'

Valid Usage

handleType must be a value included in the Handle Types Supported by VkImportFenceWin32HandleInfoKHR table.

If handleType is not VK_EXTERNAL_FENCE_HANDLE_TYPE_OPAQUE_WIN32_BIT, name must be NULL.

If handleType is not 0 and handle is NULL, name must name a valid synchronization primitive of the type specified by handleType.

If handleType is not 0 and name is NULL, handle must be a valid handle of the type specified by handleType.

If handle is not NULL, name must be NULL.

If handle is not NULL, it must obey any requirements listed for handleType in external fence handle types compatibility.

If name is not NULL, it must obey any requirements listed for handleType in external fence handle types compatibility.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMPORT_FENCE_WIN32_HANDLE_INFO_KHR

pNext must be NULL

fence must be a valid VkFence handle

flags must be a valid combination of VkFenceImportFlagBits values

If handleType is not 0, handleType must be a valid VkExternalFenceHandleTypeFlagBits value

Host Synchronization

Host access to fence must be externally synchronized

See Also

VkExternalFenceHandleTypeFlagBits, VkFence, VkFenceImportFlags, VkStructureType, vkImportFenceWin32HandleKHR

Constructors

ImportFenceWin32HandleInfoKHR

Fields

next :: Maybe SomeVkStruct

fence :: Fence

flags :: FenceImportFlags

handleType :: ExternalFenceHandleTypeFlagBits

handle :: HANDLE

name :: LPCWSTR

getFenceWin32HandleKHR :: Device -> FenceGetWin32HandleInfoKHR -> IO HANDLE

vkGetFenceWin32HandleKHR - Get a Windows HANDLE for a fence

Parameters

device is the logical device that created the fence being exported.

pGetWin32HandleInfo is a pointer to an instance of the VkFenceGetWin32HandleInfoKHR structure containing parameters of the export operation.

pHandle will return the Windows handle representing the fence state.

Description

For handle types defined as NT handles, the handles returned by vkGetFenceWin32HandleKHR are owned by the application. To avoid leaking resources, the application must release ownership of them using the CloseHandle system call when they are no longer needed.

Exporting a Windows handle from a fence may have side effects depending on the transference of the specified handle type, as described in Importing Fence Payloads.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_TOO_MANY_OBJECTS

VK_ERROR_OUT_OF_HOST_MEMORY

See Also

VkDevice, VkFenceGetWin32HandleInfoKHR

importFenceWin32HandleKHR :: Device -> ImportFenceWin32HandleInfoKHR -> IO ()

vkImportFenceWin32HandleKHR - Import a fence from a Windows HANDLE

Parameters

device is the logical device that created the fence.

pImportFenceWin32HandleInfo points to a VkImportFenceWin32HandleInfoKHR structure specifying the fence and import parameters.

Description

Importing a fence payload from Windows handles does not transfer ownership of the handle to the Vulkan implementation. For handle types defined as NT handles, the application must release ownership using the CloseHandle system call when the handle is no longer needed.

Applications can import the same fence payload into multiple instances of Vulkan, into the same instance from which it was exported, and multiple times into a given Vulkan instance.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_INVALID_EXTERNAL_HANDLE

See Also

VkDevice, VkImportFenceWin32HandleInfoKHR

pattern KHR_EXTERNAL_FENCE_WIN32_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_EXTERNAL_FENCE_WIN32_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_IMPORT_FENCE_WIN32_HANDLE_INFO_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_EXPORT_FENCE_WIN32_HANDLE_INFO_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_FENCE_GET_WIN32_HANDLE_INFO_KHR :: a ~ StructureType => a