Graphics.Vulkan.Extensions.VK_KHR_external_memory_fd

Documentation

getMemoryFdKHR :: Device -> MemoryGetFdInfoKHR -> IO ("fd" ::: Int32)

vkGetMemoryFdKHR - Get a POSIX file descriptor for a memory object

Parameters

Device is the logical device that created the device memory being exported.

pGetFdInfo is a pointer to a MemoryGetFdInfoKHR structure containing parameters of the export operation.

pFd will return a file descriptor representing the underlying resources of the device memory object.

Description

Each call to getMemoryFdKHR must create a new file descriptor and transfer ownership of it to the application. To avoid leaking resources, the application must release ownership of the file descriptor using the close system call when it is no longer needed, or by importing a Vulkan memory object from it. Where supported by the operating system, the implementation must set the file descriptor to be closed automatically when an execve system call is made.

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

Device, MemoryGetFdInfoKHR

getMemoryFdPropertiesKHR :: Device -> ExternalMemoryHandleTypeFlagBits -> ("fd" ::: Int32) -> IO MemoryFdPropertiesKHR

vkGetMemoryFdPropertiesKHR - Get Properties of External Memory File Descriptors

Parameters

Device is the logical device that will be importing fd.

handleType is the type of the handle fd.

fd is the handle which will be imported.

pMemoryFdProperties is a pointer to a MemoryFdPropertiesKHR structure in which the properties of the handle fd are returned.

Return Codes

Success

SUCCESS

Failure

ERROR_INVALID_EXTERNAL_HANDLE

See Also

Device, ExternalMemoryHandleTypeFlagBits, MemoryFdPropertiesKHR

data ImportMemoryFdInfoKHR

VkImportMemoryFdInfoKHR - import memory created on the same physical device from a file descriptor

Description

Importing memory from a file descriptor transfers ownership of the file descriptor from the application to the Vulkan implementation. The application must not perform any operations on the file descriptor after a successful import.

Applications can import the same underlying memory into multiple instances of Vulkan, into the same instance from which it was exported, and multiple times into a given Vulkan instance. In all cases, each import operation must create a distinct DeviceMemory object.

Valid Usage

If handleType is not 0, it must be supported for import, as reported by ExternalImageFormatProperties or ExternalBufferProperties.

The memory from which fd was exported must have been created on the same underlying physical device as Device.

If handleType is not 0, it must be defined as a POSIX file descriptor handle.

If handleType is not 0, fd must be a valid handle of the type specified by handleType.

The memory represented by fd must have been created from a physical device and driver that is compatible with Device and handleType, as described in https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#external-memory-handle-types-compatibility.

fd must obey any requirements listed for handleType in external memory handle types compatibility.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_IMPORT_MEMORY_FD_INFO_KHR

If handleType is not 0, handleType must be a valid ExternalMemoryHandleTypeFlagBits value

See Also

ExternalMemoryHandleTypeFlagBits, StructureType

Constructors

ImportMemoryFdInfoKHR

Fields

handleType :: ExternalMemoryHandleTypeFlagBits

handleType specifies the handle type of fd.

fd :: Int32

fd is the external handle to import.

data MemoryFdPropertiesKHR

VkMemoryFdPropertiesKHR - Properties of External Memory File Descriptors

Valid Usage (Implicit)

See Also

StructureType, getMemoryFdPropertiesKHR

Constructors

MemoryFdPropertiesKHR

Fields

memoryTypeBits :: Word32

memoryTypeBits is a bitmask containing one bit set for every memory type which the specified file descriptor can be imported as.

data MemoryGetFdInfoKHR

VkMemoryGetFdInfoKHR - Structure describing a POSIX FD semaphore export operation

Description

The properties of the file descriptor exported depend on the value of handleType. See ExternalMemoryHandleTypeFlagBits for a description of the properties of the defined external memory handle types.

Note

The size of the exported file may be larger than the size requested by MemoryAllocateInfo::allocationSize. If handleType is EXTERNAL_MEMORY_HANDLE_TYPE_DMA_BUF_BIT_EXT, then the application can query the file’s actual size with lseek(2).

Valid Usage (Implicit)

See Also

DeviceMemory, ExternalMemoryHandleTypeFlagBits, StructureType, getMemoryFdKHR

Constructors

MemoryGetFdInfoKHR

Fields

memory :: DeviceMemory

memory must be a valid DeviceMemory handle

handleType :: ExternalMemoryHandleTypeFlagBits

handleType must be a valid ExternalMemoryHandleTypeFlagBits value

type KHR_EXTERNAL_MEMORY_FD_SPEC_VERSION = 1

pattern KHR_EXTERNAL_MEMORY_FD_SPEC_VERSION :: forall a. Integral a => a

type KHR_EXTERNAL_MEMORY_FD_EXTENSION_NAME = "VK_KHR_external_memory_fd"

pattern KHR_EXTERNAL_MEMORY_FD_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a