Graphics.Vulkan.Extensions.VK_KHR_external_memory_win32

Documentation

withCStructExportMemoryWin32HandleInfoKHR :: ExportMemoryWin32HandleInfoKHR -> (VkExportMemoryWin32HandleInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkExportMemoryWin32HandleInfoKHR and marshal a ExportMemoryWin32HandleInfoKHR into it. The VkExportMemoryWin32HandleInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructExportMemoryWin32HandleInfoKHR :: VkExportMemoryWin32HandleInfoKHR -> IO ExportMemoryWin32HandleInfoKHR

A function to read a VkExportMemoryWin32HandleInfoKHR and all additional structures in the pointer chain into a ExportMemoryWin32HandleInfoKHR.

data ExportMemoryWin32HandleInfoKHR

VkExportMemoryWin32HandleInfoKHR - Structure specifying additional attributes of Windows handles exported from a memory

Description

If this structure is not present, or if pAttributes is set to NULL, default security descriptor values will be used, and child processes created by the application will not inherit the handle, as described in the MSDN documentation for “Synchronization Object Security and Access Rights”1. Further, if the structure is not present, the access rights will be

DXGI_SHARED_RESOURCE_READ | DXGI_SHARED_RESOURCE_WRITE

for handles of the following types:

VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_BIT

And

GENERIC_ALL

for handles of the following types:

VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_HEAP_BIT VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_RESOURCE_BIT

1

https://msdn.microsoft.com/en-us/library/windows/desktop/ms686670.aspx

Valid Usage

If VkExportMemoryAllocateInfo::handleTypes does not include VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT, VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_BIT, VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_HEAP_BIT, or VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_RESOURCE_BIT, VkExportMemoryWin32HandleInfoKHR must not be in the pNext chain of VkMemoryAllocateInfo.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_EXPORT_MEMORY_WIN32_HANDLE_INFO_KHR

If pAttributes is not NULL, pAttributes must be a valid pointer to a valid SECURITY_ATTRIBUTES value

See Also

VkStructureType

Constructors

ExportMemoryWin32HandleInfoKHR

Fields

next :: Maybe SomeVkStruct

attributes :: Ptr SECURITY_ATTRIBUTES

dwAccess :: DWORD

name :: LPCWSTR

withCStructImportMemoryWin32HandleInfoKHR :: ImportMemoryWin32HandleInfoKHR -> (VkImportMemoryWin32HandleInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkImportMemoryWin32HandleInfoKHR and marshal a ImportMemoryWin32HandleInfoKHR into it. The VkImportMemoryWin32HandleInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructImportMemoryWin32HandleInfoKHR :: VkImportMemoryWin32HandleInfoKHR -> IO ImportMemoryWin32HandleInfoKHR

A function to read a VkImportMemoryWin32HandleInfoKHR and all additional structures in the pointer chain into a ImportMemoryWin32HandleInfoKHR.

data ImportMemoryWin32HandleInfoKHR

VkImportMemoryWin32HandleInfoKHR - import Win32 memory created on the same physical device

Description

Importing memory objects from Windows handles does not transfer ownership of the handle to the Vulkan implementation. For handle types defined as NT handles, the application must release ownership using the CloseHandle system call when the handle is no longer needed.

Applications can import the same underlying memory into multiple instances of Vulkan, into the same instance from which it was exported, and multiple times into a given Vulkan instance. In all cases, each import operation must create a distinct VkDeviceMemory object.

Valid Usage

If handleType is not 0, it must be supported for import, as reported by VkExternalImageFormatProperties or VkExternalBufferProperties.

The memory from which handle was exported, or the memory named by name must have been created on the same underlying physical device as device.

If handleType is not 0, it must be defined as an NT handle or a global share handle.

If handleType is not VK_EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT, VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_TEXTURE_BIT, VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_HEAP_BIT, or VK_EXTERNAL_MEMORY_HANDLE_TYPE_D3D12_RESOURCE_BIT, name must be NULL.

If handleType is not 0 and handle is NULL, name must name a valid memory resource of the type specified by handleType.

If handleType is not 0 and name is NULL, handle must be a valid handle of the type specified by handleType.

if handle is not NULL, name must be NULL.

If handle is not NULL, it must obey any requirements listed for handleType in external memory handle types compatibility.

If name is not NULL, it must obey any requirements listed for handleType in external memory handle types compatibility.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_IMPORT_MEMORY_WIN32_HANDLE_INFO_KHR

If handleType is not 0, handleType must be a valid VkExternalMemoryHandleTypeFlagBits value

See Also

VkExternalMemoryHandleTypeFlagBits, VkStructureType

Constructors

ImportMemoryWin32HandleInfoKHR

Fields

next :: Maybe SomeVkStruct

handleType :: ExternalMemoryHandleTypeFlagBits

handle :: HANDLE

name :: LPCWSTR

withCStructMemoryGetWin32HandleInfoKHR :: MemoryGetWin32HandleInfoKHR -> (VkMemoryGetWin32HandleInfoKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkMemoryGetWin32HandleInfoKHR and marshal a MemoryGetWin32HandleInfoKHR into it. The VkMemoryGetWin32HandleInfoKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructMemoryGetWin32HandleInfoKHR :: VkMemoryGetWin32HandleInfoKHR -> IO MemoryGetWin32HandleInfoKHR

A function to read a VkMemoryGetWin32HandleInfoKHR and all additional structures in the pointer chain into a MemoryGetWin32HandleInfoKHR.

data MemoryGetWin32HandleInfoKHR

VkMemoryGetWin32HandleInfoKHR - Structure describing a Win32 handle semaphore export operation

Description

The properties of the handle returned depend on the value of handleType. See VkExternalMemoryHandleTypeFlagBits for a description of the properties of the defined external memory handle types.

Valid Usage

handleType must have been included in VkExportMemoryAllocateInfo::handleTypes when memory was created.

If handleType is defined as an NT handle, vkGetMemoryWin32HandleKHR must be called no more than once for each valid unique combination of memory and handleType.

handleType must be defined as an NT handle or a global share handle.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_MEMORY_GET_WIN32_HANDLE_INFO_KHR

pNext must be NULL

memory must be a valid VkDeviceMemory handle

handleType must be a valid VkExternalMemoryHandleTypeFlagBits value

See Also

VkDeviceMemory, VkExternalMemoryHandleTypeFlagBits, VkStructureType, vkGetMemoryWin32HandleKHR

Constructors

MemoryGetWin32HandleInfoKHR

Fields

next :: Maybe SomeVkStruct

memory :: DeviceMemory

handleType :: ExternalMemoryHandleTypeFlagBits

withCStructMemoryWin32HandlePropertiesKHR :: MemoryWin32HandlePropertiesKHR -> (VkMemoryWin32HandlePropertiesKHR -> IO a) -> IO a

A function to temporarily allocate memory for a VkMemoryWin32HandlePropertiesKHR and marshal a MemoryWin32HandlePropertiesKHR into it. The VkMemoryWin32HandlePropertiesKHR is only valid inside the provided computation and must not be returned out of it.

fromCStructMemoryWin32HandlePropertiesKHR :: VkMemoryWin32HandlePropertiesKHR -> IO MemoryWin32HandlePropertiesKHR

A function to read a VkMemoryWin32HandlePropertiesKHR and all additional structures in the pointer chain into a MemoryWin32HandlePropertiesKHR.

data MemoryWin32HandlePropertiesKHR

VkMemoryWin32HandlePropertiesKHR - Properties of External Memory Windows Handles

Valid Usage (Implicit)

See Also

VkStructureType, vkGetMemoryWin32HandlePropertiesKHR

Constructors

MemoryWin32HandlePropertiesKHR

Fields

next :: Maybe SomeVkStruct

memoryTypeBits :: Word32

getMemoryWin32HandleKHR :: Device -> MemoryGetWin32HandleInfoKHR -> IO HANDLE

vkGetMemoryWin32HandleKHR - Get a Windows HANDLE for a memory object

Parameters

device is the logical device that created the device memory being exported.

pGetWin32HandleInfo is a pointer to an instance of the VkMemoryGetWin32HandleInfoKHR structure containing parameters of the export operation.

pHandle will return the Windows handle representing the underlying resources of the device memory object.

Description

For handle types defined as NT handles, the handles returned by vkGetMemoryWin32HandleKHR are owned by the application. To avoid leaking resources, the application must release ownership of them using the CloseHandle system call when they are no longer needed.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_TOO_MANY_OBJECTS

VK_ERROR_OUT_OF_HOST_MEMORY

See Also

VkDevice, VkMemoryGetWin32HandleInfoKHR

getMemoryWin32HandlePropertiesKHR :: Device -> ExternalMemoryHandleTypeFlagBits -> HANDLE -> IO MemoryWin32HandlePropertiesKHR

vkGetMemoryWin32HandlePropertiesKHR - Get Properties of External Memory Win32 Handles

Parameters

device is the logical device that will be importing handle.

handleType is the type of the handle handle.

handle is the handle which will be imported.

pMemoryWin32HandleProperties will return properties of handle.

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_INVALID_EXTERNAL_HANDLE

See Also

VkDevice, VkExternalMemoryHandleTypeFlagBits, VkMemoryWin32HandlePropertiesKHR

pattern KHR_EXTERNAL_MEMORY_WIN32_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern KHR_EXTERNAL_MEMORY_WIN32_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_IMPORT_MEMORY_WIN32_HANDLE_INFO_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_EXPORT_MEMORY_WIN32_HANDLE_INFO_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_MEMORY_WIN32_HANDLE_PROPERTIES_KHR :: a ~ StructureType => a

pattern STRUCTURE_TYPE_MEMORY_GET_WIN32_HANDLE_INFO_KHR :: a ~ StructureType => a