Graphics.Vulkan.Extensions.VK_KHR_get_display_properties2

Documentation

getPhysicalDeviceDisplayProperties2KHR :: forall io. MonadIO io => PhysicalDevice -> io (Result, "properties" ::: Vector DisplayProperties2KHR)

vkGetPhysicalDeviceDisplayProperties2KHR - Query information about the available displays

Parameters

physicalDevice is a physical device.

pPropertyCount is a pointer to an integer related to the number of display devices available or queried, as described below.

pProperties is either NULL or a pointer to an array of DisplayProperties2KHR structures.

Description

getPhysicalDeviceDisplayProperties2KHR behaves similarly to getPhysicalDeviceDisplayPropertiesKHR, with the ability to return extended information via chained output structures.

Valid Usage (Implicit)

physicalDevice must be a valid PhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount DisplayProperties2KHR structures

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

DisplayProperties2KHR, PhysicalDevice

getPhysicalDeviceDisplayPlaneProperties2KHR :: forall io. MonadIO io => PhysicalDevice -> io (Result, "properties" ::: Vector DisplayPlaneProperties2KHR)

vkGetPhysicalDeviceDisplayPlaneProperties2KHR - Query information about the available display planes.

Parameters

physicalDevice is a physical device.

pPropertyCount is a pointer to an integer related to the number of display planes available or queried, as described below.

pProperties is either NULL or a pointer to an array of DisplayPlaneProperties2KHR structures.

Description

getPhysicalDeviceDisplayPlaneProperties2KHR behaves similarly to getPhysicalDeviceDisplayPlanePropertiesKHR, with the ability to return extended information via chained output structures.

Valid Usage (Implicit)

physicalDevice must be a valid PhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount DisplayPlaneProperties2KHR structures

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

DisplayPlaneProperties2KHR, PhysicalDevice

getDisplayModeProperties2KHR :: forall io. MonadIO io => PhysicalDevice -> DisplayKHR -> io (Result, "properties" ::: Vector DisplayModeProperties2KHR)

vkGetDisplayModeProperties2KHR - Query information about the available display modes.

Parameters

physicalDevice is the physical device associated with display.

display is the display to query.

pPropertyCount is a pointer to an integer related to the number of display modes available or queried, as described below.

pProperties is either NULL or a pointer to an array of DisplayModeProperties2KHR structures.

Description

getDisplayModeProperties2KHR behaves similarly to getDisplayModePropertiesKHR, with the ability to return extended information via chained output structures.

Valid Usage (Implicit)

physicalDevice must be a valid PhysicalDevice handle

display must be a valid DisplayKHR handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount DisplayModeProperties2KHR structures

display must have been created, allocated, or retrieved from physicalDevice

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

DisplayKHR, DisplayModeProperties2KHR, PhysicalDevice

getDisplayPlaneCapabilities2KHR :: forall io. MonadIO io => PhysicalDevice -> DisplayPlaneInfo2KHR -> io DisplayPlaneCapabilities2KHR

vkGetDisplayPlaneCapabilities2KHR - Query capabilities of a mode and plane combination

Parameters

physicalDevice is the physical device associated with pDisplayPlaneInfo.

pDisplayPlaneInfo is a pointer to a DisplayPlaneInfo2KHR structure describing the plane and mode.

pCapabilities is a pointer to a DisplayPlaneCapabilities2KHR structure in which the capabilities are returned.

Description

getDisplayPlaneCapabilities2KHR behaves similarly to getDisplayPlaneCapabilitiesKHR, with the ability to specify extended inputs via chained input structures, and to return extended information via chained output structures.

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

DisplayPlaneCapabilities2KHR, DisplayPlaneInfo2KHR, PhysicalDevice

data DisplayProperties2KHR

VkDisplayProperties2KHR - Structure describing an available display device

Valid Usage (Implicit)

See Also

DisplayPropertiesKHR, StructureType, getPhysicalDeviceDisplayProperties2KHR

Constructors

DisplayProperties2KHR

Fields

displayProperties :: DisplayPropertiesKHR

displayProperties is a DisplayPropertiesKHR structure.

data DisplayPlaneProperties2KHR

VkDisplayPlaneProperties2KHR - Structure describing an available display plane

Valid Usage (Implicit)

See Also

DisplayPlanePropertiesKHR, StructureType, getPhysicalDeviceDisplayPlaneProperties2KHR

Constructors

DisplayPlaneProperties2KHR

Fields

displayPlaneProperties :: DisplayPlanePropertiesKHR

displayPlaneProperties is a DisplayPlanePropertiesKHR structure.

data DisplayModeProperties2KHR

VkDisplayModeProperties2KHR - Structure describing an available display mode

Valid Usage (Implicit)

See Also

DisplayModePropertiesKHR, StructureType, getDisplayModeProperties2KHR

Constructors

DisplayModeProperties2KHR

Fields

displayModeProperties :: DisplayModePropertiesKHR

displayModeProperties is a DisplayModePropertiesKHR structure.

data DisplayPlaneInfo2KHR

VkDisplayPlaneInfo2KHR - Structure defining the intended configuration of a display plane

Description

Note

This parameter also implicitly specifies a display.

planeIndex is the plane which the application intends to use with the display.

The members of DisplayPlaneInfo2KHR correspond to the arguments to getDisplayPlaneCapabilitiesKHR, with sType and pNext added for extensibility.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_DISPLAY_PLANE_INFO_2_KHR

pNext must be NULL

mode must be a valid DisplayModeKHR handle

Host Synchronization

Host access to mode must be externally synchronized

See Also

DisplayModeKHR, StructureType, getDisplayPlaneCapabilities2KHR

Constructors

DisplayPlaneInfo2KHR

Fields

mode :: DisplayModeKHR

mode is the display mode the application intends to program when using the specified plane.

planeIndex :: Word32

data DisplayPlaneCapabilities2KHR

VkDisplayPlaneCapabilities2KHR - Structure describing the capabilities of a mode and plane combination

Valid Usage (Implicit)

See Also

DisplayPlaneCapabilitiesKHR, StructureType, getDisplayPlaneCapabilities2KHR

Constructors

DisplayPlaneCapabilities2KHR

Fields

capabilities :: DisplayPlaneCapabilitiesKHR

capabilities is a DisplayPlaneCapabilitiesKHR structure.

type KHR_GET_DISPLAY_PROPERTIES_2_SPEC_VERSION = 1

pattern KHR_GET_DISPLAY_PROPERTIES_2_SPEC_VERSION :: forall a. Integral a => a

type KHR_GET_DISPLAY_PROPERTIES_2_EXTENSION_NAME = "VK_KHR_get_display_properties2"

pattern KHR_GET_DISPLAY_PROPERTIES_2_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype DisplayKHR

VkDisplayKHR - Opaque handle to a display object

See Also

DisplayPlanePropertiesKHR, DisplayPropertiesKHR, acquireXlibDisplayEXT, createDisplayModeKHR, displayPowerControlEXT, getDisplayModeProperties2KHR, getDisplayModePropertiesKHR, getDisplayPlaneSupportedDisplaysKHR, getRandROutputDisplayEXT, registerDisplayEventEXT, releaseDisplayEXT

Constructors

DisplayKHR Word64

newtype DisplayModeKHR

VkDisplayModeKHR - Opaque handle to a display mode object

See Also

DisplayModePropertiesKHR, DisplayPlaneInfo2KHR, DisplaySurfaceCreateInfoKHR, createDisplayModeKHR, getDisplayPlaneCapabilitiesKHR

Constructors

DisplayModeKHR Word64

data DisplayPropertiesKHR

VkDisplayPropertiesKHR - Structure describing an available display device

Description

Note

For devices which have no natural value to return here, implementations should return the maximum resolution supported.

Note

Persistent presents may have higher latency, and may use less power when the screen content is updated infrequently, or when only a portion of the screen needs to be updated in most frames.

See Also

Bool32, DisplayKHR, DisplayProperties2KHR, Extent2D, SurfaceTransformFlagsKHR, getPhysicalDeviceDisplayPropertiesKHR

Constructors

DisplayPropertiesKHR

Fields

display :: DisplayKHR

display is a handle that is used to refer to the display described here. This handle will be valid for the lifetime of the Vulkan instance.

displayName :: ByteString

displayName is a pointer to a null-terminated UTF-8 string containing the name of the display. Generally, this will be the name provided by the display’s EDID. It can be NULL if no suitable name is available. If not NULL, the memory it points to must remain accessible as long as display is valid.

physicalDimensions :: Extent2D

physicalDimensions describes the physical width and height of the visible portion of the display, in millimeters.

physicalResolution :: Extent2D

physicalResolution describes the physical, native, or preferred resolution of the display.

supportedTransforms :: SurfaceTransformFlagsKHR

supportedTransforms is a bitmask of SurfaceTransformFlagBitsKHR describing which transforms are supported by this display.

planeReorderPossible :: Bool

planeReorderPossible tells whether the planes on this display can have their z order changed. If this is TRUE, the application can re-arrange the planes on this display in any order relative to each other.

persistentContent :: Bool

persistentContent tells whether the display supports self-refresh/internal buffering. If this is true, the application can submit persistent present operations on swapchains created against this display.

data DisplayPlanePropertiesKHR

VkDisplayPlanePropertiesKHR - Structure describing display plane properties

See Also

DisplayKHR, DisplayPlaneProperties2KHR, getPhysicalDeviceDisplayPlanePropertiesKHR

Constructors

DisplayPlanePropertiesKHR

Fields

currentDisplay :: DisplayKHR

currentDisplay is the handle of the display the plane is currently associated with. If the plane is not currently attached to any displays, this will be NULL_HANDLE.

currentStackIndex :: Word32

currentStackIndex is the current z-order of the plane. This will be between 0 and the value returned by getPhysicalDeviceDisplayPlanePropertiesKHR in pPropertyCount.

data DisplayModeParametersKHR

VkDisplayModeParametersKHR - Structure describing display parameters associated with a display mode

Description

Note

For example, a 60Hz display mode would report a refreshRate of 60,000.

Valid Usage

The width member of visibleRegion must be greater than 0

The height member of visibleRegion must be greater than 0

refreshRate must be greater than 0

See Also

DisplayModeCreateInfoKHR, DisplayModePropertiesKHR, Extent2D

Constructors

DisplayModeParametersKHR

Fields

visibleRegion :: Extent2D

visibleRegion is the 2D extents of the visible region.

refreshRate :: Word32

refreshRate is a uint32_t that is the number of times the display is refreshed each second multiplied by 1000.

data DisplayModePropertiesKHR

VkDisplayModePropertiesKHR - Structure describing display mode properties

See Also

DisplayModeKHR, DisplayModeParametersKHR, DisplayModeProperties2KHR, getDisplayModePropertiesKHR

Constructors

DisplayModePropertiesKHR

Fields

displayMode :: DisplayModeKHR

displayMode is a handle to the display mode described in this structure. This handle will be valid for the lifetime of the Vulkan instance.

parameters :: DisplayModeParametersKHR

parameters is a DisplayModeParametersKHR structure describing the display parameters associated with displayMode.

data DisplayPlaneCapabilitiesKHR

VkDisplayPlaneCapabilitiesKHR - Structure describing capabilities of a mode and plane combination

Description

The minimum and maximum position and extent fields describe the implementation limits, if any, as they apply to the specified display mode and plane. Vendors may support displaying a subset of a swapchain’s presentable images on the specified display plane. This is expressed by returning minSrcPosition, maxSrcPosition, minSrcExtent, and maxSrcExtent values that indicate a range of possible positions and sizes may be used to specify the region within the presentable images that source pixels will be read from when creating a swapchain on the specified display mode and plane.

Vendors may also support mapping the presentable images’ content to a subset or superset of the visible region in the specified display mode. This is expressed by returning minDstPosition, maxDstPosition, minDstExtent and maxDstExtent values that indicate a range of possible positions and sizes may be used to describe the region within the display mode that the source pixels will be mapped to.

Other vendors may support only a 1-1 mapping between pixels in the presentable images and the display mode. This may be indicated by returning (0,0) for minSrcPosition, maxSrcPosition, minDstPosition, and maxDstPosition, and (display mode width, display mode height) for minSrcExtent, maxSrcExtent, minDstExtent, and maxDstExtent.

These values indicate the limits of the implementation’s individual fields. Not all combinations of values within the offset and extent ranges returned in DisplayPlaneCapabilitiesKHR are guaranteed to be supported. Presentation requests specifying unsupported combinations may fail.

See Also

DisplayPlaneAlphaFlagsKHR, DisplayPlaneCapabilities2KHR, Extent2D, Offset2D, getDisplayPlaneCapabilitiesKHR

Constructors

DisplayPlaneCapabilitiesKHR

Fields

supportedAlpha :: DisplayPlaneAlphaFlagsKHR

supportedAlpha is a bitmask of DisplayPlaneAlphaFlagBitsKHR describing the supported alpha blending modes.

minSrcPosition :: Offset2D

minSrcPosition is the minimum source rectangle offset supported by this plane using the specified mode.

maxSrcPosition :: Offset2D

maxSrcPosition is the maximum source rectangle offset supported by this plane using the specified mode. The x and y components of maxSrcPosition must each be greater than or equal to the x and y components of minSrcPosition, respectively.

minSrcExtent :: Extent2D

minSrcExtent is the minimum source rectangle size supported by this plane using the specified mode.

maxSrcExtent :: Extent2D

maxSrcExtent is the maximum source rectangle size supported by this plane using the specified mode.

minDstPosition :: Offset2D

minDstPosition, maxDstPosition, minDstExtent, maxDstExtent all have similar semantics to their corresponding *Src* equivalents, but apply to the output region within the mode rather than the input region within the source image. Unlike the *Src* offsets, minDstPosition and maxDstPosition may contain negative values.

maxDstPosition :: Offset2D

minDstExtent :: Extent2D

maxDstExtent :: Extent2D

newtype DisplayPlaneAlphaFlagBitsKHR

VkDisplayPlaneAlphaFlagBitsKHR - Alpha blending type

See Also

DisplayPlaneAlphaFlagsKHR, DisplaySurfaceCreateInfoKHR

Constructors

DisplayPlaneAlphaFlagBitsKHR Flags

Bundled Patterns

pattern DISPLAY_PLANE_ALPHA_OPAQUE_BIT_KHR :: DisplayPlaneAlphaFlagBitsKHR
DISPLAY_PLANE_ALPHA_OPAQUE_BIT_KHR specifies that the source image will be treated as opaque.

pattern DISPLAY_PLANE_ALPHA_GLOBAL_BIT_KHR :: DisplayPlaneAlphaFlagBitsKHR
DISPLAY_PLANE_ALPHA_GLOBAL_BIT_KHR specifies that a global alpha value must be specified that will be applied to all pixels in the source image.

pattern DISPLAY_PLANE_ALPHA_PER_PIXEL_BIT_KHR :: DisplayPlaneAlphaFlagBitsKHR
DISPLAY_PLANE_ALPHA_PER_PIXEL_BIT_KHR specifies that the alpha value will be determined by the alpha channel of the source image’s pixels. If the source format contains no alpha values, no blending will be applied. The source alpha values are not premultiplied into the source image’s other color channels.

pattern DISPLAY_PLANE_ALPHA_PER_PIXEL_PREMULTIPLIED_BIT_KHR :: DisplayPlaneAlphaFlagBitsKHR
DISPLAY_PLANE_ALPHA_PER_PIXEL_PREMULTIPLIED_BIT_KHR is equivalent to DISPLAY_PLANE_ALPHA_PER_PIXEL_BIT_KHR, except the source alpha values are assumed to be premultiplied into the source image’s other color channels.

type DisplayPlaneAlphaFlagsKHR = DisplayPlaneAlphaFlagBitsKHR

newtype SurfaceTransformFlagBitsKHR

VkSurfaceTransformFlagBitsKHR - presentation transforms supported on a device

See Also

CommandBufferInheritanceRenderPassTransformInfoQCOM, DisplaySurfaceCreateInfoKHR, RenderPassTransformBeginInfoQCOM, SurfaceCapabilities2EXT, SurfaceCapabilitiesKHR, SurfaceTransformFlagsKHR, SwapchainCreateInfoKHR

Constructors

SurfaceTransformFlagBitsKHR Flags

Bundled Patterns

pattern SURFACE_TRANSFORM_IDENTITY_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_IDENTITY_BIT_KHR specifies that image content is presented without being transformed.

pattern SURFACE_TRANSFORM_ROTATE_90_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_ROTATE_90_BIT_KHR specifies that image content is rotated 90 degrees clockwise.

pattern SURFACE_TRANSFORM_ROTATE_180_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_ROTATE_180_BIT_KHR specifies that image content is rotated 180 degrees clockwise.

pattern SURFACE_TRANSFORM_ROTATE_270_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_ROTATE_270_BIT_KHR specifies that image content is rotated 270 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_BIT_KHR specifies that image content is mirrored horizontally.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_90_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_90_BIT_KHR specifies that image content is mirrored horizontally, then rotated 90 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_180_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_180_BIT_KHR specifies that image content is mirrored horizontally, then rotated 180 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_270_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_270_BIT_KHR specifies that image content is mirrored horizontally, then rotated 270 degrees clockwise.

pattern SURFACE_TRANSFORM_INHERIT_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_INHERIT_BIT_KHR specifies that the presentation transform is not specified, and is instead determined by platform-specific considerations and mechanisms outside Vulkan.

type SurfaceTransformFlagsKHR = SurfaceTransformFlagBitsKHR