Graphics.Vulkan.Extensions.VK_KHR_xcb_surface

Documentation

createXcbSurfaceKHR :: Instance -> XcbSurfaceCreateInfoKHR -> ("allocator" ::: Maybe AllocationCallbacks) -> IO SurfaceKHR

vkCreateXcbSurfaceKHR - Create a SurfaceKHR object for a X11 window, using the XCB client-side library

Parameters

Instance is the instance to associate the surface with.

pCreateInfo is a pointer to a XcbSurfaceCreateInfoKHR structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface is a pointer to a SurfaceKHR handle in which the created surface object is returned.

Valid Usage (Implicit)

Instance must be a valid Instance handle

pCreateInfo must be a valid pointer to a valid XcbSurfaceCreateInfoKHR structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pSurface must be a valid pointer to a SurfaceKHR handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, Instance, SurfaceKHR, XcbSurfaceCreateInfoKHR

getPhysicalDeviceXcbPresentationSupportKHR :: PhysicalDevice -> ("queueFamilyIndex" ::: Word32) -> Ptr Xcb_connection_t -> ("visual_id" ::: Xcb_visualid_t) -> IO Bool

vkGetPhysicalDeviceXcbPresentationSupportKHR - Query physical device for presentation to X11 server using XCB

Parameters

PhysicalDevice is the physical device.

queueFamilyIndex is the queue family index.

connection is a pointer to an Xcb_connection_t to the X server. visual_id is an X11 visual (Xcb_visualid_t).

Description

This platform-specific function can be called prior to creating a surface.

Valid Usage (Implicit)

See Also

PhysicalDevice

data XcbSurfaceCreateInfoKHR

VkXcbSurfaceCreateInfoKHR - Structure specifying parameters of a newly created Xcb surface object

Valid Usage (Implicit)

See Also

StructureType, XcbSurfaceCreateFlagsKHR, createXcbSurfaceKHR

Constructors

XcbSurfaceCreateInfoKHR

Fields

flags :: XcbSurfaceCreateFlagsKHR

Flags must be 0

connection :: Ptr Xcb_connection_t

connection must point to a valid X11 Xcb_connection_t.

window :: Xcb_window_t

Window must be a valid X11 Xcb_window_t.

newtype XcbSurfaceCreateFlagsKHR

Constructors

XcbSurfaceCreateFlagsKHR Flags

type KHR_XCB_SURFACE_SPEC_VERSION = 6

pattern KHR_XCB_SURFACE_SPEC_VERSION :: forall a. Integral a => a

type KHR_XCB_SURFACE_EXTENSION_NAME = "VK_KHR_xcb_surface"

pattern KHR_XCB_SURFACE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype SurfaceKHR

VkSurfaceKHR - Opaque handle to a surface object

Description

The VK_KHR_surface extension declares the SurfaceKHR object, and provides a function for destroying SurfaceKHR objects. Separate platform-specific extensions each provide a function for creating a SurfaceKHR object for the respective platform. From the application’s perspective this is an opaque handle, just like the handles of other Vulkan objects.

See Also

PhysicalDeviceSurfaceInfo2KHR, SwapchainCreateInfoKHR, createAndroidSurfaceKHR, createDisplayPlaneSurfaceKHR, createHeadlessSurfaceEXT, createIOSSurfaceMVK, createImagePipeSurfaceFUCHSIA, createMacOSSurfaceMVK, createMetalSurfaceEXT, createStreamDescriptorSurfaceGGP, createViSurfaceNN, createWaylandSurfaceKHR, createWin32SurfaceKHR, createXcbSurfaceKHR, createXlibSurfaceKHR, destroySurfaceKHR, getDeviceGroupSurfacePresentModesKHR, getPhysicalDevicePresentRectanglesKHR, getPhysicalDeviceSurfaceCapabilities2EXT, getPhysicalDeviceSurfaceCapabilitiesKHR, getPhysicalDeviceSurfaceFormatsKHR, getPhysicalDeviceSurfacePresentModesKHR, getPhysicalDeviceSurfaceSupportKHR

Constructors

SurfaceKHR Word64

type Xcb_visualid_t = Word32

type Xcb_window_t = Word32

data Xcb_connection_t