Graphics.Vulkan.Extensions.VK_NN_vi_surface

Documentation

createViSurfaceNN :: forall io. MonadIO io => Instance -> ViSurfaceCreateInfoNN -> ("allocator" ::: Maybe AllocationCallbacks) -> io SurfaceKHR

vkCreateViSurfaceNN - Create a SurfaceKHR object for a VI layer

Parameters

instance is the instance with which to associate the surface.

pCreateInfo is a pointer to a ViSurfaceCreateInfoNN structure containing parameters affecting the creation of the surface object.

pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

pSurface is a pointer to a SurfaceKHR handle in which the created surface object is returned.

Description

During the lifetime of a surface created using a particular nn::vi::NativeWindowHandle, applications must not attempt to create another surface for the same nn::vi::Layer or attempt to connect to the same nn::vi::Layer through other platform mechanisms.

If the native window is created with a specified size, currentExtent will reflect that size. In this case, applications should use the same size for the swapchain’s imageExtent. Otherwise, the currentExtent will have the special value (0xFFFFFFFF, 0xFFFFFFFF), indicating that applications are expected to choose an appropriate size for the swapchain’s imageExtent (e.g., by matching the result of a call to nn::vi::GetDisplayResolution).

Valid Usage (Implicit)

instance must be a valid Instance handle

pCreateInfo must be a valid pointer to a valid ViSurfaceCreateInfoNN structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pSurface must be a valid pointer to a SurfaceKHR handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_NATIVE_WINDOW_IN_USE_KHR

See Also

AllocationCallbacks, Instance, SurfaceKHR, ViSurfaceCreateInfoNN

data ViSurfaceCreateInfoNN

VkViSurfaceCreateInfoNN - Structure specifying parameters of a newly created VI surface object

Valid Usage (Implicit)

See Also

StructureType, ViSurfaceCreateFlagsNN, createViSurfaceNN

Constructors

ViSurfaceCreateInfoNN

Fields

flags :: ViSurfaceCreateFlagsNN

flags must be 0

window :: Ptr ()

window must be a valid nn::vi::NativeWindowHandle

newtype ViSurfaceCreateFlagsNN

Constructors

ViSurfaceCreateFlagsNN Flags

type NN_VI_SURFACE_SPEC_VERSION = 1

pattern NN_VI_SURFACE_SPEC_VERSION :: forall a. Integral a => a

type NN_VI_SURFACE_EXTENSION_NAME = "VK_NN_vi_surface"

pattern NN_VI_SURFACE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype SurfaceKHR

VkSurfaceKHR - Opaque handle to a surface object

Description

The VK_KHR_surface extension declares the SurfaceKHR object, and provides a function for destroying SurfaceKHR objects. Separate platform-specific extensions each provide a function for creating a SurfaceKHR object for the respective platform. From the application’s perspective this is an opaque handle, just like the handles of other Vulkan objects.

See Also

PhysicalDeviceSurfaceInfo2KHR, SwapchainCreateInfoKHR, createAndroidSurfaceKHR, createDisplayPlaneSurfaceKHR, createHeadlessSurfaceEXT, createIOSSurfaceMVK, createImagePipeSurfaceFUCHSIA, createMacOSSurfaceMVK, createMetalSurfaceEXT, createStreamDescriptorSurfaceGGP, createViSurfaceNN, createWaylandSurfaceKHR, createWin32SurfaceKHR, createXcbSurfaceKHR, createXlibSurfaceKHR, destroySurfaceKHR, getDeviceGroupSurfacePresentModesKHR, getPhysicalDevicePresentRectanglesKHR, getPhysicalDeviceSurfaceCapabilities2EXT, getPhysicalDeviceSurfaceCapabilitiesKHR, getPhysicalDeviceSurfaceFormatsKHR, getPhysicalDeviceSurfacePresentModesKHR, getPhysicalDeviceSurfaceSupportKHR

Constructors

SurfaceKHR Word64