Graphics.Vulkan.Extensions.VK_NVX_device_generated_commands

Documentation

withCStructCmdProcessCommandsInfoNVX :: CmdProcessCommandsInfoNVX -> (VkCmdProcessCommandsInfoNVX -> IO a) -> IO a

A function to temporarily allocate memory for a VkCmdProcessCommandsInfoNVX and marshal a CmdProcessCommandsInfoNVX into it. The VkCmdProcessCommandsInfoNVX is only valid inside the provided computation and must not be returned out of it.

fromCStructCmdProcessCommandsInfoNVX :: DeviceCmds -> VkCmdProcessCommandsInfoNVX -> IO CmdProcessCommandsInfoNVX

A function to read a VkCmdProcessCommandsInfoNVX and all additional structures in the pointer chain into a CmdProcessCommandsInfoNVX.

data CmdProcessCommandsInfoNVX

VkCmdProcessCommandsInfoNVX - Structure specifying parameters for the generation of commands

Valid Usage

The provided objectTable must include all objects referenced by the generation process

indirectCommandsTokenCount must match the indirectCommandsLayout’s tokenCount

The tokenType member of each entry in the pIndirectCommandsTokens array must match the values used at creation time of indirectCommandsLayout

If targetCommandBuffer is provided, it must have reserved command space

If targetCommandBuffer is provided, the objectTable must match the reservation’s objectTable and must have had all referenced objects registered at reservation time

If targetCommandBuffer is provided, the indirectCommandsLayout must match the reservation’s indirectCommandsLayout

If targetCommandBuffer is provided, the maxSequencesCount must not exceed the reservation’s maxSequencesCount

If sequencesCountBuffer is used, its usage flag must have the VK_BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set

If sequencesCountBuffer is used, sequencesCountOffset must be aligned to VkDeviceGeneratedCommandsLimitsNVX::minSequenceCountBufferOffsetAlignment

If sequencesIndexBuffer is used, its usage flag must have the VK_BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set

If sequencesIndexBuffer is used, sequencesIndexOffset must be aligned to VkDeviceGeneratedCommandsLimitsNVX::minSequenceIndexBufferOffsetAlignment

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_CMD_PROCESS_COMMANDS_INFO_NVX

pNext must be NULL

objectTable must be a valid VkObjectTableNVX handle

indirectCommandsLayout must be a valid VkIndirectCommandsLayoutNVX handle

pIndirectCommandsTokens must be a valid pointer to an array of indirectCommandsTokenCount valid VkIndirectCommandsTokenNVX structures

If targetCommandBuffer is not NULL, targetCommandBuffer must be a valid VkCommandBuffer handle

If sequencesCountBuffer is not VK_NULL_HANDLE, sequencesCountBuffer must be a valid VkBuffer handle

If sequencesIndexBuffer is not VK_NULL_HANDLE, sequencesIndexBuffer must be a valid VkBuffer handle

indirectCommandsTokenCount must be greater than 0

Each of indirectCommandsLayout, objectTable, sequencesCountBuffer, sequencesIndexBuffer, and targetCommandBuffer that are valid handles must have been created, allocated, or retrieved from the same VkDevice

Host Synchronization

Host access to objectTable must be externally synchronized

Host access to targetCommandBuffer must be externally synchronized

See Also

VkBuffer, VkCommandBuffer, VkDeviceSize, VkIndirectCommandsLayoutNVX, VkIndirectCommandsTokenNVX, VkObjectTableNVX, VkStructureType, vkCmdProcessCommandsNVX

Constructors

CmdProcessCommandsInfoNVX

Fields

next :: Maybe SomeVkStruct

objectTable :: ObjectTableNVX

indirectCommandsLayout :: IndirectCommandsLayoutNVX

indirectCommandsTokens :: Vector IndirectCommandsTokenNVX

maxSequencesCount :: Word32

targetCommandBuffer :: CommandBuffer

sequencesCountBuffer :: Buffer

sequencesCountOffset :: DeviceSize

sequencesIndexBuffer :: Buffer

sequencesIndexOffset :: DeviceSize

withCStructCmdReserveSpaceForCommandsInfoNVX :: CmdReserveSpaceForCommandsInfoNVX -> (VkCmdReserveSpaceForCommandsInfoNVX -> IO a) -> IO a

A function to temporarily allocate memory for a VkCmdReserveSpaceForCommandsInfoNVX and marshal a CmdReserveSpaceForCommandsInfoNVX into it. The VkCmdReserveSpaceForCommandsInfoNVX is only valid inside the provided computation and must not be returned out of it.

fromCStructCmdReserveSpaceForCommandsInfoNVX :: VkCmdReserveSpaceForCommandsInfoNVX -> IO CmdReserveSpaceForCommandsInfoNVX

A function to read a VkCmdReserveSpaceForCommandsInfoNVX and all additional structures in the pointer chain into a CmdReserveSpaceForCommandsInfoNVX.

data CmdReserveSpaceForCommandsInfoNVX

VkCmdReserveSpaceForCommandsInfoNVX - Structure specifying parameters for the reservation of command buffer space

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_CMD_RESERVE_SPACE_FOR_COMMANDS_INFO_NVX

pNext must be NULL

objectTable must be a valid VkObjectTableNVX handle

indirectCommandsLayout must be a valid VkIndirectCommandsLayoutNVX handle

Both of indirectCommandsLayout, and objectTable must have been created, allocated, or retrieved from the same VkDevice

Host Synchronization

Host access to objectTable must be externally synchronized

See Also

VkIndirectCommandsLayoutNVX, VkObjectTableNVX, VkStructureType, vkCmdReserveSpaceForCommandsNVX

Constructors

CmdReserveSpaceForCommandsInfoNVX

Fields

next :: Maybe SomeVkStruct

objectTable :: ObjectTableNVX

indirectCommandsLayout :: IndirectCommandsLayoutNVX

maxSequencesCount :: Word32

withCStructDeviceGeneratedCommandsFeaturesNVX :: DeviceGeneratedCommandsFeaturesNVX -> (VkDeviceGeneratedCommandsFeaturesNVX -> IO a) -> IO a

A function to temporarily allocate memory for a VkDeviceGeneratedCommandsFeaturesNVX and marshal a DeviceGeneratedCommandsFeaturesNVX into it. The VkDeviceGeneratedCommandsFeaturesNVX is only valid inside the provided computation and must not be returned out of it.

fromCStructDeviceGeneratedCommandsFeaturesNVX :: VkDeviceGeneratedCommandsFeaturesNVX -> IO DeviceGeneratedCommandsFeaturesNVX

A function to read a VkDeviceGeneratedCommandsFeaturesNVX and all additional structures in the pointer chain into a DeviceGeneratedCommandsFeaturesNVX.

data DeviceGeneratedCommandsFeaturesNVX

VkDeviceGeneratedCommandsFeaturesNVX - Structure specifying physical device support

Valid Usage (Implicit)

See Also

VkBool32, VkStructureType, vkGetPhysicalDeviceGeneratedCommandsPropertiesNVX

Constructors

DeviceGeneratedCommandsFeaturesNVX

Fields

next :: Maybe SomeVkStruct

computeBindingPointSupport :: Bool

withCStructDeviceGeneratedCommandsLimitsNVX :: DeviceGeneratedCommandsLimitsNVX -> (VkDeviceGeneratedCommandsLimitsNVX -> IO a) -> IO a

A function to temporarily allocate memory for a VkDeviceGeneratedCommandsLimitsNVX and marshal a DeviceGeneratedCommandsLimitsNVX into it. The VkDeviceGeneratedCommandsLimitsNVX is only valid inside the provided computation and must not be returned out of it.

fromCStructDeviceGeneratedCommandsLimitsNVX :: VkDeviceGeneratedCommandsLimitsNVX -> IO DeviceGeneratedCommandsLimitsNVX

A function to read a VkDeviceGeneratedCommandsLimitsNVX and all additional structures in the pointer chain into a DeviceGeneratedCommandsLimitsNVX.

data DeviceGeneratedCommandsLimitsNVX

VkDeviceGeneratedCommandsLimitsNVX - Structure specifying physical device limits

Valid Usage (Implicit)

See Also

VkStructureType, vkGetPhysicalDeviceGeneratedCommandsPropertiesNVX

Constructors

DeviceGeneratedCommandsLimitsNVX

Fields

next :: Maybe SomeVkStruct

maxIndirectCommandsLayoutTokenCount :: Word32

maxObjectEntryCounts :: Word32

minSequenceCountBufferOffsetAlignment :: Word32

minSequenceIndexBufferOffsetAlignment :: Word32

minCommandsTokenBufferOffsetAlignment :: Word32

withCStructIndirectCommandsLayoutCreateInfoNVX :: IndirectCommandsLayoutCreateInfoNVX -> (VkIndirectCommandsLayoutCreateInfoNVX -> IO a) -> IO a

A function to temporarily allocate memory for a VkIndirectCommandsLayoutCreateInfoNVX and marshal a IndirectCommandsLayoutCreateInfoNVX into it. The VkIndirectCommandsLayoutCreateInfoNVX is only valid inside the provided computation and must not be returned out of it.

fromCStructIndirectCommandsLayoutCreateInfoNVX :: VkIndirectCommandsLayoutCreateInfoNVX -> IO IndirectCommandsLayoutCreateInfoNVX

A function to read a VkIndirectCommandsLayoutCreateInfoNVX and all additional structures in the pointer chain into a IndirectCommandsLayoutCreateInfoNVX.

data IndirectCommandsLayoutCreateInfoNVX

VkIndirectCommandsLayoutCreateInfoNVX - Structure specifying the parameters of a newly created indirect commands layout object

Description

The following code illustrates some of the key flags:

void cmdProcessAllSequences(cmd, objectTable, indirectCommandsLayout, pIndirectCommandsTokens, sequencesCount, indexbuffer, indexbufferoffset)
{
  for (s = 0; s < sequencesCount; s++)
  {
    sequence = s;

    if (indirectCommandsLayout.flags & VK_INDIRECT_COMMANDS_LAYOUT_USAGE_UNORDERED_SEQUENCES_BIT_NVX) {
      sequence = incoherent_implementation_dependent_permutation[ sequence ];
    }
    if (indirectCommandsLayout.flags & VK_INDIRECT_COMMANDS_LAYOUT_USAGE_INDEXED_SEQUENCES_BIT_NVX) {
      sequence = indexbuffer.load_uint32( sequence * sizeof(uint32_t) + indexbufferoffset);
    }

    cmdProcessSequence( cmd, objectTable, indirectCommandsLayout, pIndirectCommandsTokens, sequence );
  }
}

Valid Usage

tokenCount must be greater than 0 and below VkDeviceGeneratedCommandsLimitsNVX::maxIndirectCommandsLayoutTokenCount

If the VkDeviceGeneratedCommandsFeaturesNVX::computeBindingPointSupport feature is not enabled, then pipelineBindPoint must not be VK_PIPELINE_BIND_POINT_COMPUTE

If pTokens contains an entry of VK_INDIRECT_COMMANDS_TOKEN_TYPE_PIPELINE_NVX it must be the first element of the array and there must be only a single element of such token type.

All state binding tokens in pTokens must occur prior work provoking tokens (VK_INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_NVX, VK_INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_INDEXED_NVX, VK_INDIRECT_COMMANDS_TOKEN_TYPE_DISPATCH_NVX).

The content of pTokens must include one single work provoking token that is compatible with the pipelineBindPoint.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_INDIRECT_COMMANDS_LAYOUT_CREATE_INFO_NVX

pNext must be NULL

pipelineBindPoint must be a valid VkPipelineBindPoint value

flags must be a valid combination of VkIndirectCommandsLayoutUsageFlagBitsNVX values

flags must not be 0

pTokens must be a valid pointer to an array of tokenCount valid VkIndirectCommandsLayoutTokenNVX structures

tokenCount must be greater than 0

See Also

VkIndirectCommandsLayoutTokenNVX, VkIndirectCommandsLayoutUsageFlagsNVX, VkPipelineBindPoint, VkStructureType, vkCreateIndirectCommandsLayoutNVX

Constructors

IndirectCommandsLayoutCreateInfoNVX

Fields

next :: Maybe SomeVkStruct

pipelineBindPoint :: PipelineBindPoint

flags :: IndirectCommandsLayoutUsageFlagsNVX

tokens :: Vector IndirectCommandsLayoutTokenNVX

type IndirectCommandsLayoutNVX = VkIndirectCommandsLayoutNVX

VkIndirectCommandsLayoutNVX - Opaque handle to an indirect commands layout object

See Also

VkCmdProcessCommandsInfoNVX, VkCmdReserveSpaceForCommandsInfoNVX, vkCreateIndirectCommandsLayoutNVX, vkDestroyIndirectCommandsLayoutNVX

withCStructIndirectCommandsLayoutTokenNVX :: IndirectCommandsLayoutTokenNVX -> (VkIndirectCommandsLayoutTokenNVX -> IO a) -> IO a

A function to temporarily allocate memory for a VkIndirectCommandsLayoutTokenNVX and marshal a IndirectCommandsLayoutTokenNVX into it. The VkIndirectCommandsLayoutTokenNVX is only valid inside the provided computation and must not be returned out of it.

fromCStructIndirectCommandsLayoutTokenNVX :: VkIndirectCommandsLayoutTokenNVX -> IO IndirectCommandsLayoutTokenNVX

A function to read a VkIndirectCommandsLayoutTokenNVX and all additional structures in the pointer chain into a IndirectCommandsLayoutTokenNVX.

data IndirectCommandsLayoutTokenNVX

VkIndirectCommandsLayoutTokenNVX - Struct specifying the details of an indirect command layout token

Valid Usage (Implicit)

See Also

VkIndirectCommandsLayoutCreateInfoNVX, VkIndirectCommandsTokenTypeNVX

Constructors

IndirectCommandsLayoutTokenNVX

Fields

tokenType :: IndirectCommandsTokenTypeNVX

bindingUnit :: Word32

dynamicCount :: Word32

divisor :: Word32

type IndirectCommandsLayoutUsageFlagBitsNVX = VkIndirectCommandsLayoutUsageFlagBitsNVX

VkIndirectCommandsLayoutUsageFlagBitsNVX - Bitmask specifying allowed usage of an indirect commands layout

See Also

VkIndirectCommandsLayoutUsageFlagsNVX

pattern INDIRECT_COMMANDS_LAYOUT_USAGE_UNORDERED_SEQUENCES_BIT_NVX :: a ~ IndirectCommandsLayoutUsageFlagBitsNVX => a

VK_INDIRECT_COMMANDS_LAYOUT_USAGE_UNORDERED_SEQUENCES_BIT_NVX specifies that the processing of sequences can happen at an implementation-dependent order, which is not guaranteed to be coherent across multiple invocations.

pattern INDIRECT_COMMANDS_LAYOUT_USAGE_SPARSE_SEQUENCES_BIT_NVX :: a ~ IndirectCommandsLayoutUsageFlagBitsNVX => a

VK_INDIRECT_COMMANDS_LAYOUT_USAGE_SPARSE_SEQUENCES_BIT_NVX specifies that there is likely a high difference between allocated number of sequences and actually used.

pattern INDIRECT_COMMANDS_LAYOUT_USAGE_EMPTY_EXECUTIONS_BIT_NVX :: a ~ IndirectCommandsLayoutUsageFlagBitsNVX => a

VK_INDIRECT_COMMANDS_LAYOUT_USAGE_EMPTY_EXECUTIONS_BIT_NVX specifies that there are likely many draw or dispatch calls that are zero-sized (zero grid dimension, no primitives to render).

pattern INDIRECT_COMMANDS_LAYOUT_USAGE_INDEXED_SEQUENCES_BIT_NVX :: a ~ IndirectCommandsLayoutUsageFlagBitsNVX => a

VK_INDIRECT_COMMANDS_LAYOUT_USAGE_INDEXED_SEQUENCES_BIT_NVX specifies that the input data for the sequences is not implicitly indexed from 0..sequencesUsed but a user provided VkBuffer encoding the index is provided.

type IndirectCommandsLayoutUsageFlagsNVX = IndirectCommandsLayoutUsageFlagBitsNVX

VkIndirectCommandsLayoutUsageFlagsNVX - Bitmask of VkIndirectCommandsLayoutUsageFlagBitsNVX

Description

VkIndirectCommandsLayoutUsageFlagsNVX is a bitmask type for setting a mask of zero or more VkIndirectCommandsLayoutUsageFlagBitsNVX.

See Also

VkIndirectCommandsLayoutCreateInfoNVX, VkIndirectCommandsLayoutUsageFlagBitsNVX

withCStructIndirectCommandsTokenNVX :: IndirectCommandsTokenNVX -> (VkIndirectCommandsTokenNVX -> IO a) -> IO a

A function to temporarily allocate memory for a VkIndirectCommandsTokenNVX and marshal a IndirectCommandsTokenNVX into it. The VkIndirectCommandsTokenNVX is only valid inside the provided computation and must not be returned out of it.

fromCStructIndirectCommandsTokenNVX :: VkIndirectCommandsTokenNVX -> IO IndirectCommandsTokenNVX

A function to read a VkIndirectCommandsTokenNVX and all additional structures in the pointer chain into a IndirectCommandsTokenNVX.

data IndirectCommandsTokenNVX

VkIndirectCommandsTokenNVX - Structure specifying parameters for the reservation of command buffer space

Valid Usage

The buffer’s usage flag must have the VK_BUFFER_USAGE_INDIRECT_BUFFER_BIT bit set.

The offset must be aligned to VkDeviceGeneratedCommandsLimitsNVX::minCommandsTokenBufferOffsetAlignment.

Valid Usage (Implicit)

tokenType must be a valid VkIndirectCommandsTokenTypeNVX value

buffer must be a valid VkBuffer handle

See Also

VkBuffer, VkCmdProcessCommandsInfoNVX, VkDeviceSize, VkIndirectCommandsTokenTypeNVX

Constructors

IndirectCommandsTokenNVX

Fields

tokenType :: IndirectCommandsTokenTypeNVX

buffer :: Buffer

offset :: DeviceSize

type IndirectCommandsTokenTypeNVX = VkIndirectCommandsTokenTypeNVX

VkIndirectCommandsTokenTypeNVX - Enum specifying

Description

'

+-----------------------------------------------+----------------------+
| Token type                                    | Equivalent command   |
+===============================================+======================+
| 'Graphics.Vulkan.C.Extensions.VK_NVX_device_g | 'Graphics.Vulkan.C.C |
| enerated_commands.VK_INDIRECT_COMMANDS_TOKEN_ | ore10.CommandBufferB |
| TYPE_PIPELINE_NVX'                            | uilding.vkCmdBindPip |
|                                               | eline'               |
+-----------------------------------------------+----------------------+
| 'Graphics.Vulkan.C.Extensions.VK_NVX_device_g | 'Graphics.Vulkan.C.C |
| enerated_commands.VK_INDIRECT_COMMANDS_TOKEN_ | ore10.CommandBufferB |
| TYPE_DESCRIPTOR_SET_NVX'                      | uilding.vkCmdBindDes |
|                                               | criptorSets'         |
+-----------------------------------------------+----------------------+
| 'Graphics.Vulkan.C.Extensions.VK_NVX_device_g | 'Graphics.Vulkan.C.C |
| enerated_commands.VK_INDIRECT_COMMANDS_TOKEN_ | ore10.CommandBufferB |
| TYPE_INDEX_BUFFER_NVX'                        | uilding.vkCmdBindInd |
|                                               | exBuffer'            |
+-----------------------------------------------+----------------------+
| 'Graphics.Vulkan.C.Extensions.VK_NVX_device_g | 'Graphics.Vulkan.C.C |
| enerated_commands.VK_INDIRECT_COMMANDS_TOKEN_ | ore10.CommandBufferB |
| TYPE_VERTEX_BUFFER_NVX'                       | uilding.vkCmdBindVer |
|                                               | texBuffers'          |
+-----------------------------------------------+----------------------+
| 'Graphics.Vulkan.C.Extensions.VK_NVX_device_g | 'Graphics.Vulkan.C.C |
| enerated_commands.VK_INDIRECT_COMMANDS_TOKEN_ | ore10.CommandBufferB |
| TYPE_PUSH_CONSTANT_NVX'                       | uilding.vkCmdPushCon |
|                                               | stants'              |
+-----------------------------------------------+----------------------+
| 'Graphics.Vulkan.C.Extensions.VK_NVX_device_g | 'Graphics.Vulkan.C.C |
| enerated_commands.VK_INDIRECT_COMMANDS_TOKEN_ | ore10.CommandBufferB |
| TYPE_DRAW_INDEXED_NVX'                        | uilding.vkCmdDrawInd |
|                                               | exedIndirect'        |
+-----------------------------------------------+----------------------+
| 'Graphics.Vulkan.C.Extensions.VK_NVX_device_g | 'Graphics.Vulkan.C.C |
| enerated_commands.VK_INDIRECT_COMMANDS_TOKEN_ | ore10.CommandBufferB |
| TYPE_DRAW_NVX'                                | uilding.vkCmdDrawInd |
|                                               | irect'               |
+-----------------------------------------------+----------------------+
| 'Graphics.Vulkan.C.Extensions.VK_NVX_device_g | 'Graphics.Vulkan.C.C |
| enerated_commands.VK_INDIRECT_COMMANDS_TOKEN_ | ore10.CommandBufferB |
| TYPE_DISPATCH_NVX'                            | uilding.vkCmdDispatc |
|                                               | hIndirect'           |
+-----------------------------------------------+----------------------+

Supported indirect command tokens

See Also

VkIndirectCommandsLayoutTokenNVX, VkIndirectCommandsTokenNVX

pattern INDIRECT_COMMANDS_TOKEN_TYPE_PIPELINE_NVX :: a ~ IndirectCommandsTokenTypeNVX => a

pattern INDIRECT_COMMANDS_TOKEN_TYPE_DESCRIPTOR_SET_NVX :: a ~ IndirectCommandsTokenTypeNVX => a

pattern INDIRECT_COMMANDS_TOKEN_TYPE_INDEX_BUFFER_NVX :: a ~ IndirectCommandsTokenTypeNVX => a

pattern INDIRECT_COMMANDS_TOKEN_TYPE_VERTEX_BUFFER_NVX :: a ~ IndirectCommandsTokenTypeNVX => a

pattern INDIRECT_COMMANDS_TOKEN_TYPE_PUSH_CONSTANT_NVX :: a ~ IndirectCommandsTokenTypeNVX => a

pattern INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_INDEXED_NVX :: a ~ IndirectCommandsTokenTypeNVX => a

pattern INDIRECT_COMMANDS_TOKEN_TYPE_DRAW_NVX :: a ~ IndirectCommandsTokenTypeNVX => a

pattern INDIRECT_COMMANDS_TOKEN_TYPE_DISPATCH_NVX :: a ~ IndirectCommandsTokenTypeNVX => a

type ObjectEntryTypeNVX = VkObjectEntryTypeNVX

VkObjectEntryTypeNVX - Enum specifying object table entry type

See Also

VkObjectTableCreateInfoNVX, VkObjectTableDescriptorSetEntryNVX, VkObjectTableEntryNVX, VkObjectTableIndexBufferEntryNVX, VkObjectTablePipelineEntryNVX, VkObjectTablePushConstantEntryNVX, VkObjectTableVertexBufferEntryNVX, vkUnregisterObjectsNVX

pattern OBJECT_ENTRY_TYPE_DESCRIPTOR_SET_NVX :: a ~ ObjectEntryTypeNVX => a

VK_OBJECT_ENTRY_TYPE_DESCRIPTOR_SET_NVX specifies a VkDescriptorSet resource entry that is registered via VkObjectTableDescriptorSetEntryNVX.

pattern OBJECT_ENTRY_TYPE_PIPELINE_NVX :: a ~ ObjectEntryTypeNVX => a

VK_OBJECT_ENTRY_TYPE_PIPELINE_NVX specifies a VkPipeline resource entry that is registered via VkObjectTablePipelineEntryNVX.

pattern OBJECT_ENTRY_TYPE_INDEX_BUFFER_NVX :: a ~ ObjectEntryTypeNVX => a

VK_OBJECT_ENTRY_TYPE_INDEX_BUFFER_NVX specifies a VkBuffer resource entry that is registered via VkObjectTableIndexBufferEntryNVX.

pattern OBJECT_ENTRY_TYPE_VERTEX_BUFFER_NVX :: a ~ ObjectEntryTypeNVX => a

VK_OBJECT_ENTRY_TYPE_VERTEX_BUFFER_NVX specifies a VkBuffer resource entry that is registered via VkObjectTableVertexBufferEntryNVX.

pattern OBJECT_ENTRY_TYPE_PUSH_CONSTANT_NVX :: a ~ ObjectEntryTypeNVX => a

VK_OBJECT_ENTRY_TYPE_PUSH_CONSTANT_NVX specifies the resource entry is registered via VkObjectTablePushConstantEntryNVX.

type ObjectEntryUsageFlagBitsNVX = VkObjectEntryUsageFlagBitsNVX

VkObjectEntryUsageFlagBitsNVX - Bitmask specifying allowed usage of an object entry

See Also

VkObjectEntryUsageFlagsNVX

pattern OBJECT_ENTRY_USAGE_GRAPHICS_BIT_NVX :: a ~ ObjectEntryUsageFlagBitsNVX => a

VK_OBJECT_ENTRY_USAGE_GRAPHICS_BIT_NVX specifies that the resource is bound to VK_PIPELINE_BIND_POINT_GRAPHICS

pattern OBJECT_ENTRY_USAGE_COMPUTE_BIT_NVX :: a ~ ObjectEntryUsageFlagBitsNVX => a

VK_OBJECT_ENTRY_USAGE_COMPUTE_BIT_NVX specifies that the resource is bound to VK_PIPELINE_BIND_POINT_COMPUTE

type ObjectEntryUsageFlagsNVX = ObjectEntryUsageFlagBitsNVX

VkObjectEntryUsageFlagsNVX - Bitmask of VkObjectEntryUsageFlagBitsNVX

Description

VkObjectEntryUsageFlagsNVX is a bitmask type for setting a mask of zero or more VkObjectEntryUsageFlagBitsNVX.

See Also

VkObjectEntryUsageFlagBitsNVX, VkObjectTableCreateInfoNVX, VkObjectTableDescriptorSetEntryNVX, VkObjectTableEntryNVX, VkObjectTableIndexBufferEntryNVX, VkObjectTablePipelineEntryNVX, VkObjectTablePushConstantEntryNVX, VkObjectTableVertexBufferEntryNVX

withCStructObjectTableCreateInfoNVX :: ObjectTableCreateInfoNVX -> (VkObjectTableCreateInfoNVX -> IO a) -> IO a

A function to temporarily allocate memory for a VkObjectTableCreateInfoNVX and marshal a ObjectTableCreateInfoNVX into it. The VkObjectTableCreateInfoNVX is only valid inside the provided computation and must not be returned out of it.

fromCStructObjectTableCreateInfoNVX :: VkObjectTableCreateInfoNVX -> IO ObjectTableCreateInfoNVX

A function to read a VkObjectTableCreateInfoNVX and all additional structures in the pointer chain into a ObjectTableCreateInfoNVX.

data ObjectTableCreateInfoNVX

VkObjectTableCreateInfoNVX - Structure specifying the parameters of a newly created object table

Valid Usage

If the VkDeviceGeneratedCommandsFeaturesNVX::computeBindingPointSupport feature is not enabled, pObjectEntryUsageFlags must not contain VK_OBJECT_ENTRY_USAGE_COMPUTE_BIT_NVX

Any value within pObjectEntryCounts must not exceed VkDeviceGeneratedCommandsLimitsNVX::maxObjectEntryCounts

maxUniformBuffersPerDescriptor must be within the limits supported by the device.

maxStorageBuffersPerDescriptor must be within the limits supported by the device.

maxStorageImagesPerDescriptor must be within the limits supported by the device.

maxSampledImagesPerDescriptor must be within the limits supported by the device.

Valid Usage (Implicit)

sType must be VK_STRUCTURE_TYPE_OBJECT_TABLE_CREATE_INFO_NVX

pNext must be NULL

pObjectEntryTypes must be a valid pointer to an array of objectCount valid VkObjectEntryTypeNVX values

pObjectEntryCounts must be a valid pointer to an array of objectCount uint32_t values

pObjectEntryUsageFlags must be a valid pointer to an array of objectCount valid combinations of VkObjectEntryUsageFlagBitsNVX values

Each element of pObjectEntryUsageFlags must not be 0

objectCount must be greater than 0

See Also

VkObjectEntryTypeNVX, VkObjectEntryUsageFlagsNVX, VkStructureType, vkCreateObjectTableNVX

Constructors

ObjectTableCreateInfoNVX

Fields

next :: Maybe SomeVkStruct

objectEntryTypes :: Vector ObjectEntryTypeNVX

objectEntryCounts :: Vector Word32

objectEntryUsageFlags :: Vector ObjectEntryUsageFlagsNVX

maxUniformBuffersPerDescriptor :: Word32

maxStorageBuffersPerDescriptor :: Word32

maxStorageImagesPerDescriptor :: Word32

maxSampledImagesPerDescriptor :: Word32

maxPipelineLayouts :: Word32

withCStructObjectTableDescriptorSetEntryNVX :: ObjectTableDescriptorSetEntryNVX -> (VkObjectTableDescriptorSetEntryNVX -> IO a) -> IO a

A function to temporarily allocate memory for a VkObjectTableDescriptorSetEntryNVX and marshal a ObjectTableDescriptorSetEntryNVX into it. The VkObjectTableDescriptorSetEntryNVX is only valid inside the provided computation and must not be returned out of it.

fromCStructObjectTableDescriptorSetEntryNVX :: VkObjectTableDescriptorSetEntryNVX -> IO ObjectTableDescriptorSetEntryNVX

A function to read a VkObjectTableDescriptorSetEntryNVX and all additional structures in the pointer chain into a ObjectTableDescriptorSetEntryNVX.

data ObjectTableDescriptorSetEntryNVX

VkObjectTableDescriptorSetEntryNVX - Parameters of an object table descriptor set entry

Valid Usage

type must be VK_OBJECT_ENTRY_TYPE_DESCRIPTOR_SET_NVX

Valid Usage (Implicit)

type must be a valid VkObjectEntryTypeNVX value

flags must be a valid combination of VkObjectEntryUsageFlagBitsNVX values

flags must not be 0

pipelineLayout must be a valid VkPipelineLayout handle

descriptorSet must be a valid VkDescriptorSet handle

Both of descriptorSet, and pipelineLayout must have been created, allocated, or retrieved from the same VkDevice

See Also

VkDescriptorSet, VkObjectEntryTypeNVX, VkObjectEntryUsageFlagsNVX, VkPipelineLayout

Constructors

ObjectTableDescriptorSetEntryNVX

Fields

type' :: ObjectEntryTypeNVX

flags :: ObjectEntryUsageFlagsNVX

pipelineLayout :: PipelineLayout

descriptorSet :: DescriptorSet

withCStructObjectTableEntryNVX :: ObjectTableEntryNVX -> (VkObjectTableEntryNVX -> IO a) -> IO a

A function to temporarily allocate memory for a VkObjectTableEntryNVX and marshal a ObjectTableEntryNVX into it. The VkObjectTableEntryNVX is only valid inside the provided computation and must not be returned out of it.

fromCStructObjectTableEntryNVX :: VkObjectTableEntryNVX -> IO ObjectTableEntryNVX

A function to read a VkObjectTableEntryNVX and all additional structures in the pointer chain into a ObjectTableEntryNVX.

data ObjectTableEntryNVX

VkObjectTableEntryNVX - Common parameters of an object table resource entry

Valid Usage

If the VkDeviceGeneratedCommandsFeaturesNVX::computeBindingPointSupport feature is not enabled, flags must not contain VK_OBJECT_ENTRY_USAGE_COMPUTE_BIT_NVX

Valid Usage (Implicit)

type must be a valid VkObjectEntryTypeNVX value

flags must be a valid combination of VkObjectEntryUsageFlagBitsNVX values

flags must not be 0

See Also

VkObjectEntryTypeNVX, VkObjectEntryUsageFlagsNVX, vkRegisterObjectsNVX

Constructors

ObjectTableEntryNVX

Fields

type' :: ObjectEntryTypeNVX

flags :: ObjectEntryUsageFlagsNVX

withCStructObjectTableIndexBufferEntryNVX :: ObjectTableIndexBufferEntryNVX -> (VkObjectTableIndexBufferEntryNVX -> IO a) -> IO a

A function to temporarily allocate memory for a VkObjectTableIndexBufferEntryNVX and marshal a ObjectTableIndexBufferEntryNVX into it. The VkObjectTableIndexBufferEntryNVX is only valid inside the provided computation and must not be returned out of it.

fromCStructObjectTableIndexBufferEntryNVX :: VkObjectTableIndexBufferEntryNVX -> IO ObjectTableIndexBufferEntryNVX

A function to read a VkObjectTableIndexBufferEntryNVX and all additional structures in the pointer chain into a ObjectTableIndexBufferEntryNVX.

data ObjectTableIndexBufferEntryNVX

VkObjectTableIndexBufferEntryNVX - Parameters of an object table index buffer entry

Valid Usage (Implicit)

See Also

VkBuffer, VkIndexType, VkObjectEntryTypeNVX, VkObjectEntryUsageFlagsNVX

Constructors

ObjectTableIndexBufferEntryNVX

Fields

type' :: ObjectEntryTypeNVX

flags :: ObjectEntryUsageFlagsNVX

buffer :: Buffer

indexType :: IndexType

type ObjectTableNVX = VkObjectTableNVX

VkObjectTableNVX - Opaque handle to an object table

See Also

VkCmdProcessCommandsInfoNVX, VkCmdReserveSpaceForCommandsInfoNVX, vkCreateObjectTableNVX, vkDestroyObjectTableNVX, vkRegisterObjectsNVX, vkUnregisterObjectsNVX

withCStructObjectTablePipelineEntryNVX :: ObjectTablePipelineEntryNVX -> (VkObjectTablePipelineEntryNVX -> IO a) -> IO a

A function to temporarily allocate memory for a VkObjectTablePipelineEntryNVX and marshal a ObjectTablePipelineEntryNVX into it. The VkObjectTablePipelineEntryNVX is only valid inside the provided computation and must not be returned out of it.

fromCStructObjectTablePipelineEntryNVX :: VkObjectTablePipelineEntryNVX -> IO ObjectTablePipelineEntryNVX

A function to read a VkObjectTablePipelineEntryNVX and all additional structures in the pointer chain into a ObjectTablePipelineEntryNVX.

data ObjectTablePipelineEntryNVX

VkObjectTablePipelineEntryNVX - Parameters of an object table pipeline entry

Valid Usage (Implicit)

See Also

VkObjectEntryTypeNVX, VkObjectEntryUsageFlagsNVX, VkPipeline

Constructors

ObjectTablePipelineEntryNVX

Fields

type' :: ObjectEntryTypeNVX

flags :: ObjectEntryUsageFlagsNVX

pipeline :: Pipeline

withCStructObjectTablePushConstantEntryNVX :: ObjectTablePushConstantEntryNVX -> (VkObjectTablePushConstantEntryNVX -> IO a) -> IO a

A function to temporarily allocate memory for a VkObjectTablePushConstantEntryNVX and marshal a ObjectTablePushConstantEntryNVX into it. The VkObjectTablePushConstantEntryNVX is only valid inside the provided computation and must not be returned out of it.

fromCStructObjectTablePushConstantEntryNVX :: VkObjectTablePushConstantEntryNVX -> IO ObjectTablePushConstantEntryNVX

A function to read a VkObjectTablePushConstantEntryNVX and all additional structures in the pointer chain into a ObjectTablePushConstantEntryNVX.

data ObjectTablePushConstantEntryNVX

VkObjectTablePushConstantEntryNVX - Parameters of an object table push constant entry

Valid Usage (Implicit)

See Also

VkObjectEntryTypeNVX, VkObjectEntryUsageFlagsNVX, VkPipelineLayout, VkShaderStageFlags

Constructors

ObjectTablePushConstantEntryNVX

Fields

type' :: ObjectEntryTypeNVX

flags :: ObjectEntryUsageFlagsNVX

pipelineLayout :: PipelineLayout

stageFlags :: ShaderStageFlags

withCStructObjectTableVertexBufferEntryNVX :: ObjectTableVertexBufferEntryNVX -> (VkObjectTableVertexBufferEntryNVX -> IO a) -> IO a

A function to temporarily allocate memory for a VkObjectTableVertexBufferEntryNVX and marshal a ObjectTableVertexBufferEntryNVX into it. The VkObjectTableVertexBufferEntryNVX is only valid inside the provided computation and must not be returned out of it.

fromCStructObjectTableVertexBufferEntryNVX :: VkObjectTableVertexBufferEntryNVX -> IO ObjectTableVertexBufferEntryNVX

A function to read a VkObjectTableVertexBufferEntryNVX and all additional structures in the pointer chain into a ObjectTableVertexBufferEntryNVX.

data ObjectTableVertexBufferEntryNVX

VkObjectTableVertexBufferEntryNVX - Parameters of an object table vertex buffer entry

Valid Usage (Implicit)

See Also

VkBuffer, VkObjectEntryTypeNVX, VkObjectEntryUsageFlagsNVX

Constructors

ObjectTableVertexBufferEntryNVX

Fields

type' :: ObjectEntryTypeNVX

flags :: ObjectEntryUsageFlagsNVX

buffer :: Buffer

cmdProcessCommandsNVX :: CommandBuffer -> CmdProcessCommandsInfoNVX -> IO ()

vkCmdProcessCommandsNVX - Performs the generation of commands on the device

Parameters

commandBuffer is the primary command buffer in which the generation process takes space.

pProcessCommandsInfo is a pointer to an instance of the VkCmdProcessCommandsInfoNVX structure containing parameters affecting the processing of commands.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pProcessCommandsInfo must be a valid pointer to a valid VkCmdProcessCommandsInfoNVX structure

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

This command must only be called inside of a render pass instance

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Primary         | Inside          | Graphics        |                 |
| Secondary       |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCmdProcessCommandsInfoNVX, VkCommandBuffer

cmdReserveSpaceForCommandsNVX :: CommandBuffer -> CmdReserveSpaceForCommandsInfoNVX -> IO ()

vkCmdReserveSpaceForCommandsNVX - Perform a reservation of command buffer space

Parameters

commandBuffer is the secondary command buffer in which the space for device-generated commands is reserved.

pProcessCommandsInfo is a pointer to an instance of the VkCmdReserveSpaceForCommandsInfoNVX structure containing parameters affecting the reservation of command buffer space.

Valid Usage

The provided commandBuffer must not have had a prior space reservation since its creation or the last reset.

The state of the commandBuffer must be legal to execute all commands within the sequence provided by the indirectCommandsLayout member of pProcessCommandsInfo.

Valid Usage (Implicit)

commandBuffer must be a valid VkCommandBuffer handle

pReserveSpaceInfo must be a valid pointer to a valid VkCmdReserveSpaceForCommandsInfoNVX structure

commandBuffer must be in the recording state

The VkCommandPool that commandBuffer was allocated from must support graphics, or compute operations

This command must only be called inside of a render pass instance

commandBuffer must be a secondary VkCommandBuffer

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the VkCommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

+-----------------+-----------------+-----------------+-----------------+
| <https://www.kh | <https://www.kh | <https://www.kh | <https://www.kh |
| ronos.org/regis | ronos.org/regis | ronos.org/regis | ronos.org/regis |
| try/vulkan/spec | try/vulkan/spec | try/vulkan/spec | try/vulkan/spec |
| s/1.0-extension | s/1.0-extension | s/1.0-extension | s/1.0-extension |
| s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h | s/html/vkspec.h |
| tml#VkCommandBu | tml#vkCmdBeginR | tml#VkQueueFlag | tml#synchroniza |
| fferLevel Comma | enderPass Rende | Bits Supported  | tion-pipeline-s |
| nd Buffer Level | r Pass Scope>   | Queue Types>    | tages-types Pip |
| s>              |                 |                 | eline Type>     |
+=================+=================+=================+=================+
| Secondary       | Inside          | Graphics        |                 |
|                 |                 | Compute         |                 |
+-----------------+-----------------+-----------------+-----------------+

See Also

VkCmdReserveSpaceForCommandsInfoNVX, VkCommandBuffer

createIndirectCommandsLayoutNVX :: Device -> IndirectCommandsLayoutCreateInfoNVX -> Maybe AllocationCallbacks -> IO IndirectCommandsLayoutNVX

vkCreateIndirectCommandsLayoutNVX - Create an indirect command layout object

Parameters

device is the logical device that creates the indirect command layout.

pCreateInfo is a pointer to an instance of the VkIndirectCommandsLayoutCreateInfoNVX structure containing parameters affecting creation of the indirect command layout.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pIndirectCommandsLayout points to a VkIndirectCommandsLayoutNVX handle in which the resulting indirect command layout is returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkIndirectCommandsLayoutCreateInfoNVX structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pIndirectCommandsLayout must be a valid pointer to a VkIndirectCommandsLayoutNVX handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDevice, VkIndirectCommandsLayoutCreateInfoNVX, VkIndirectCommandsLayoutNVX

createObjectTableNVX :: Device -> ObjectTableCreateInfoNVX -> Maybe AllocationCallbacks -> IO ObjectTableNVX

vkCreateObjectTableNVX - Create an object table

Parameters

device is the logical device that creates the object table.

pCreateInfo is a pointer to an instance of the VkObjectTableCreateInfoNVX structure containing parameters affecting creation of the table.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pObjectTable points to a VkObjectTableNVX handle in which the resulting object table is returned.

Valid Usage (Implicit)

device must be a valid VkDevice handle

pCreateInfo must be a valid pointer to a valid VkObjectTableCreateInfoNVX structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

pObjectTable must be a valid pointer to a VkObjectTableNVX handle

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkAllocationCallbacks, VkDevice, VkObjectTableCreateInfoNVX, VkObjectTableNVX

destroyIndirectCommandsLayoutNVX :: Device -> IndirectCommandsLayoutNVX -> Maybe AllocationCallbacks -> IO ()

vkDestroyIndirectCommandsLayoutNVX - Destroy an object table

Parameters

device is the logical device that destroys the layout.

indirectCommandsLayout is the table to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to indirectCommandsLayout must have completed execution

If VkAllocationCallbacks were provided when objectTable was created, a compatible set of callbacks must be provided here

If no VkAllocationCallbacks were provided when objectTable was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid VkDevice handle

indirectCommandsLayout must be a valid VkIndirectCommandsLayoutNVX handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

indirectCommandsLayout must have been created, allocated, or retrieved from device

See Also

VkAllocationCallbacks, VkDevice, VkIndirectCommandsLayoutNVX

destroyObjectTableNVX :: Device -> ObjectTableNVX -> Maybe AllocationCallbacks -> IO ()

vkDestroyObjectTableNVX - Destroy an object table

Parameters

device is the logical device that destroys the table.

objectTable is the table to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to objectTable must have completed execution.

If VkAllocationCallbacks were provided when objectTable was created, a compatible set of callbacks must be provided here.

If no VkAllocationCallbacks were provided when objectTable was created, pAllocator must be NULL.

Valid Usage (Implicit)

device must be a valid VkDevice handle

objectTable must be a valid VkObjectTableNVX handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid VkAllocationCallbacks structure

objectTable must have been created, allocated, or retrieved from device

Host Synchronization

Host access to objectTable must be externally synchronized

See Also

VkAllocationCallbacks, VkDevice, VkObjectTableNVX

getPhysicalDeviceGeneratedCommandsPropertiesNVX :: PhysicalDevice -> IO (DeviceGeneratedCommandsFeaturesNVX, DeviceGeneratedCommandsLimitsNVX)

vkGetPhysicalDeviceGeneratedCommandsPropertiesNVX - Returns device-generated commands related properties of a physical device

Parameters

physicalDevice is the handle to the physical device whose properties will be queried.

pFeatures points to an instance of the VkDeviceGeneratedCommandsFeaturesNVX structure, that will be filled with returned information.

pLimits points to an instance of the VkDeviceGeneratedCommandsLimitsNVX structure, that will be filled with returned information.

Valid Usage (Implicit)

See Also

VkDeviceGeneratedCommandsFeaturesNVX, VkDeviceGeneratedCommandsLimitsNVX, VkPhysicalDevice

registerObjectsNVX :: Device -> ObjectTableNVX -> Vector ObjectTableEntryNVX -> Vector Word32 -> IO ()

vkRegisterObjectsNVX - Register resource bindings in an object table

Parameters

device is the logical device that creates the object table.

objectTable is the table for which the resources are registered.

objectCount is the number of resources to register.

ppObjectTableEntries provides an array for detailed binding informations, each array element is a pointer to a struct of type VkObjectTablePipelineEntryNVX, VkObjectTableDescriptorSetEntryNVX, VkObjectTableVertexBufferEntryNVX, VkObjectTableIndexBufferEntryNVX or VkObjectTablePushConstantEntryNVX (see below for details).

pObjectIndices are the indices at which each resource is registered.

Valid Usage

The contents of pObjectTableEntry must yield plausible bindings supported by the device.

At any pObjectIndices there must not be a registered resource already.

Any value inside pObjectIndices must be below the appropriate VkObjectTableCreateInfoNVX::pObjectEntryCounts limits provided at objectTable creation time.

Valid Usage (Implicit)

device must be a valid VkDevice handle

objectTable must be a valid VkObjectTableNVX handle

ppObjectTableEntries must be a valid pointer to an array of objectCount valid VkObjectTableEntryNVX structures

pObjectIndices must be a valid pointer to an array of objectCount uint32_t values

objectCount must be greater than 0

objectTable must have been created, allocated, or retrieved from device

Host Synchronization

Host access to objectTable must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDevice, VkObjectTableEntryNVX, VkObjectTableNVX

unregisterObjectsNVX :: Device -> ObjectTableNVX -> Vector ObjectEntryTypeNVX -> Vector Word32 -> IO ()

vkUnregisterObjectsNVX - Unregister resource bindings in an object table

Parameters

device is the logical device that creates the object table.

objectTable is the table from which the resources are unregistered.

objectCount is the number of resources being removed from the object table.

pObjectEntryType provides an array of VkObjectEntryTypeNVX for the resources being removed.

pObjectIndices provides the array of object indices to be removed.

Valid Usage

At any pObjectIndices there must be a registered resource already.

The pObjectEntryTypes of the resource at pObjectIndices must match.

All operations on the device using the registered resource must have been completed.

Valid Usage (Implicit)

device must be a valid VkDevice handle

objectTable must be a valid VkObjectTableNVX handle

pObjectEntryTypes must be a valid pointer to an array of objectCount valid VkObjectEntryTypeNVX values

pObjectIndices must be a valid pointer to an array of objectCount uint32_t values

objectCount must be greater than 0

objectTable must have been created, allocated, or retrieved from device

Host Synchronization

Host access to objectTable must be externally synchronized

Return Codes

Success

- VK_SUCCESS

Failure

- VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

See Also

VkDevice, VkObjectEntryTypeNVX, VkObjectTableNVX

withIndirectCommandsLayoutNVX :: Device -> IndirectCommandsLayoutCreateInfoNVX -> Maybe AllocationCallbacks -> (IndirectCommandsLayoutNVX -> IO a) -> IO a

A safe wrapper for createIndirectCommandsLayoutNVX and destroyIndirectCommandsLayoutNVX using bracket

The allocated value must not be returned from the provided computation

withObjectTableNVX :: Device -> ObjectTableCreateInfoNVX -> Maybe AllocationCallbacks -> (ObjectTableNVX -> IO a) -> IO a

A safe wrapper for createObjectTableNVX and destroyObjectTableNVX using bracket

The allocated value must not be returned from the provided computation

withRegisteredObjectsNVX :: Device -> ObjectTableNVX -> Vector ObjectTableEntryNVX -> Vector Word32 -> Vector ObjectEntryTypeNVX -> IO a -> IO a

A safe wrapper for registerObjectsNVX and unregisterObjectsNVX using bracket_

The allocated value must not be returned from the provided computation

pattern NVX_DEVICE_GENERATED_COMMANDS_EXTENSION_NAME :: (Eq a, IsString a) => a

pattern NVX_DEVICE_GENERATED_COMMANDS_SPEC_VERSION :: Integral a => a

pattern STRUCTURE_TYPE_OBJECT_TABLE_CREATE_INFO_NVX :: a ~ StructureType => a

pattern STRUCTURE_TYPE_INDIRECT_COMMANDS_LAYOUT_CREATE_INFO_NVX :: a ~ StructureType => a

pattern STRUCTURE_TYPE_CMD_PROCESS_COMMANDS_INFO_NVX :: a ~ StructureType => a

pattern STRUCTURE_TYPE_CMD_RESERVE_SPACE_FOR_COMMANDS_INFO_NVX :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEVICE_GENERATED_COMMANDS_LIMITS_NVX :: a ~ StructureType => a

pattern STRUCTURE_TYPE_DEVICE_GENERATED_COMMANDS_FEATURES_NVX :: a ~ StructureType => a

pattern PIPELINE_STAGE_COMMAND_PROCESS_BIT_NVX :: a ~ PipelineStageFlagBits => a

pattern ACCESS_COMMAND_PROCESS_READ_BIT_NVX :: a ~ AccessFlagBits => a

pattern ACCESS_COMMAND_PROCESS_WRITE_BIT_NVX :: a ~ AccessFlagBits => a

pattern OBJECT_TYPE_OBJECT_TABLE_NVX :: a ~ ObjectType => a

pattern OBJECT_TYPE_INDIRECT_COMMANDS_LAYOUT_NVX :: a ~ ObjectType => a