Graphics.Vulkan.Extensions.VK_NV_cooperative_matrix

Documentation

getPhysicalDeviceCooperativeMatrixPropertiesNV :: forall io. MonadIO io => PhysicalDevice -> io (Result, "properties" ::: Vector CooperativeMatrixPropertiesNV)

vkGetPhysicalDeviceCooperativeMatrixPropertiesNV - Returns properties describing what cooperative matrix types are supported

Parameters

physicalDevice is the physical device.

pPropertyCount is a pointer to an integer related to the number of cooperative matrix properties available or queried.

pProperties is either NULL or a pointer to an array of CooperativeMatrixPropertiesNV structures.

Description

If pProperties is NULL, then the number of cooperative matrix properties available is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If pPropertyCount is less than the number of cooperative matrix properties available, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of cooperative matrix properties available, INCOMPLETE will be returned instead of SUCCESS, to indicate that not all the available cooperative matrix properties were returned.

Valid Usage (Implicit)

physicalDevice must be a valid PhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount CooperativeMatrixPropertiesNV structures

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

CooperativeMatrixPropertiesNV, PhysicalDevice

data PhysicalDeviceCooperativeMatrixFeaturesNV

VkPhysicalDeviceCooperativeMatrixFeaturesNV - Structure describing cooperative matrix features that can be supported by an implementation

Members

The members of the PhysicalDeviceCooperativeMatrixFeaturesNV structure describe the following features:

Description

If the PhysicalDeviceCooperativeMatrixFeaturesNV structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether the feature is supported. PhysicalDeviceCooperativeMatrixFeaturesNV can also be included in the pNext chain of DeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceCooperativeMatrixFeaturesNV

Fields

cooperativeMatrix :: Bool

cooperativeMatrix indicates that the implementation supports the CooperativeMatrixNV SPIR-V capability.

cooperativeMatrixRobustBufferAccess :: Bool

cooperativeMatrixRobustBufferAccess indicates that the implementation supports robust buffer access for SPIR-V OpCooperativeMatrixLoadNV and OpCooperativeMatrixStoreNV instructions.

data PhysicalDeviceCooperativeMatrixPropertiesNV

VkPhysicalDeviceCooperativeMatrixPropertiesNV - Structure describing cooperative matrix properties supported by an implementation

Members

The members of the PhysicalDeviceCooperativeMatrixPropertiesNV structure describe the following implementation-dependent limits:

Description

If the PhysicalDeviceCooperativeMatrixPropertiesNV structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

ShaderStageFlags, StructureType

Constructors

PhysicalDeviceCooperativeMatrixPropertiesNV

Fields

cooperativeMatrixSupportedStages :: ShaderStageFlags

cooperativeMatrixSupportedStages is a bitfield of ShaderStageFlagBits describing the shader stages that cooperative matrix instructions are supported in. cooperativeMatrixSupportedStages will have the SHADER_STAGE_COMPUTE_BIT bit set if any of the physical device’s queues support QUEUE_COMPUTE_BIT.

data CooperativeMatrixPropertiesNV

VkCooperativeMatrixPropertiesNV - Structure specifying cooperative matrix properties

Description

If some types are preferred over other types (e.g. for performance), they should appear earlier in the list enumerated by getPhysicalDeviceCooperativeMatrixPropertiesNV.

At least one entry in the list must have power of two values for all of MSize, KSize, and NSize.

Valid Usage (Implicit)

See Also

ComponentTypeNV, ScopeNV, StructureType, getPhysicalDeviceCooperativeMatrixPropertiesNV

Constructors

CooperativeMatrixPropertiesNV

Fields

mSize :: Word32

MSize is the number of rows in matrices A, C, and D.

nSize :: Word32

NSize is the number of columns in matrices B, C, D.

kSize :: Word32

KSize is the number of columns in matrix A and rows in matrix B.

aType :: ComponentTypeNV

AType must be a valid ComponentTypeNV value

bType :: ComponentTypeNV

BType must be a valid ComponentTypeNV value

cType :: ComponentTypeNV

CType must be a valid ComponentTypeNV value

dType :: ComponentTypeNV

DType must be a valid ComponentTypeNV value

scope :: ScopeNV

scope must be a valid ScopeNV value

newtype ScopeNV

VkScopeNV - Specify SPIR-V scope

Description

All enum values match the corresponding SPIR-V value.

See Also

CooperativeMatrixPropertiesNV

Constructors

ScopeNV Int32

Bundled Patterns

pattern SCOPE_DEVICE_NV :: ScopeNV
SCOPE_DEVICE_NV corresponds to SPIR-V Device scope.

pattern SCOPE_WORKGROUP_NV :: ScopeNV
SCOPE_WORKGROUP_NV corresponds to SPIR-V Workgroup scope.

pattern SCOPE_SUBGROUP_NV :: ScopeNV
SCOPE_SUBGROUP_NV corresponds to SPIR-V Subgroup scope.

pattern SCOPE_QUEUE_FAMILY_NV :: ScopeNV
SCOPE_QUEUE_FAMILY_NV corresponds to SPIR-V QueueFamily scope.

newtype ComponentTypeNV

VkComponentTypeNV - Specify SPIR-V cooperative matrix component type

See Also

CooperativeMatrixPropertiesNV

Constructors

ComponentTypeNV Int32

Bundled Patterns

pattern COMPONENT_TYPE_FLOAT16_NV :: ComponentTypeNV
COMPONENT_TYPE_FLOAT16_NV corresponds to SPIR-V OpTypeFloat 16.

pattern COMPONENT_TYPE_FLOAT32_NV :: ComponentTypeNV
COMPONENT_TYPE_FLOAT32_NV corresponds to SPIR-V OpTypeFloat 32.

pattern COMPONENT_TYPE_FLOAT64_NV :: ComponentTypeNV
COMPONENT_TYPE_FLOAT64_NV corresponds to SPIR-V OpTypeFloat 64.

pattern COMPONENT_TYPE_SINT8_NV :: ComponentTypeNV
COMPONENT_TYPE_SINT8_NV corresponds to SPIR-V OpTypeInt 8 1.

pattern COMPONENT_TYPE_SINT16_NV :: ComponentTypeNV
COMPONENT_TYPE_SINT16_NV corresponds to SPIR-V OpTypeInt 16 1.

pattern COMPONENT_TYPE_SINT32_NV :: ComponentTypeNV
COMPONENT_TYPE_SINT32_NV corresponds to SPIR-V OpTypeInt 32 1.

pattern COMPONENT_TYPE_SINT64_NV :: ComponentTypeNV
COMPONENT_TYPE_SINT64_NV corresponds to SPIR-V OpTypeInt 64 1.

pattern COMPONENT_TYPE_UINT8_NV :: ComponentTypeNV
COMPONENT_TYPE_UINT8_NV corresponds to SPIR-V OpTypeInt 8 0.

pattern COMPONENT_TYPE_UINT16_NV :: ComponentTypeNV
COMPONENT_TYPE_UINT16_NV corresponds to SPIR-V OpTypeInt 16 0.

pattern COMPONENT_TYPE_UINT32_NV :: ComponentTypeNV
COMPONENT_TYPE_UINT32_NV corresponds to SPIR-V OpTypeInt 32 0.

pattern COMPONENT_TYPE_UINT64_NV :: ComponentTypeNV
COMPONENT_TYPE_UINT64_NV corresponds to SPIR-V OpTypeInt 64 0.

type NV_COOPERATIVE_MATRIX_SPEC_VERSION = 1

pattern NV_COOPERATIVE_MATRIX_SPEC_VERSION :: forall a. Integral a => a

type NV_COOPERATIVE_MATRIX_EXTENSION_NAME = "VK_NV_cooperative_matrix"

pattern NV_COOPERATIVE_MATRIX_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a