Graphics.Vulkan.Extensions.VK_NV_fragment_coverage_to_color

Documentation

data PipelineCoverageToColorStateCreateInfoNV

VkPipelineCoverageToColorStateCreateInfoNV - Structure specifying whether fragment coverage replaces a color

Description

If coverageToColorEnable is TRUE, the fragment coverage information is treated as a bitmask with one bit for each sample (as in the Sample Mask section), and this bitmask replaces the first component of the color value corresponding to the fragment shader output location with Location equal to coverageToColorLocation and Index equal to zero. If the color attachment format has fewer bits than the sample coverage, the low bits of the sample coverage bitmask are taken without any clamping. If the color attachment format has more bits than the sample coverage, the high bits of the sample coverage bitmask are filled with zeros.

If Sample Shading is in use, the coverage bitmask only has bits set for samples that correspond to the fragment shader invocation that shades those samples.

This pipeline stage occurs after sample counting and before blending, and is always performed after fragment shading regardless of the setting of EarlyFragmentTests.

If coverageToColorEnable is FALSE, these operations are skipped. If this structure is not present, it is as if coverageToColorEnable is FALSE.

Valid Usage

If coverageToColorEnable is TRUE, then the render pass subpass indicated by 'Graphics.Vulkan.Core10.Pipeline.GraphicsPipelineCreateInfo'::'Graphics.Vulkan.Core10.Handles.RenderPass' and GraphicsPipelineCreateInfo::subpass must have a color attachment at the location selected by coverageToColorLocation, with a Format of FORMAT_R8_UINT, FORMAT_R8_SINT, FORMAT_R16_UINT, FORMAT_R16_SINT, FORMAT_R32_UINT, or FORMAT_R32_SINT

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PIPELINE_COVERAGE_TO_COLOR_STATE_CREATE_INFO_NV

Flags must be 0

See Also

Bool32, PipelineCoverageToColorStateCreateFlagsNV, StructureType

Constructors

PipelineCoverageToColorStateCreateInfoNV

Fields

flags :: PipelineCoverageToColorStateCreateFlagsNV

Flags is reserved for future use.

coverageToColorEnable :: Bool

coverageToColorEnable controls whether the fragment coverage value replaces a fragment color output.

coverageToColorLocation :: Word32

coverageToColorLocation controls which fragment shader color output value is replaced.

newtype PipelineCoverageToColorStateCreateFlagsNV

VkPipelineCoverageToColorStateCreateFlagsNV - Reserved for future use

Description

PipelineCoverageToColorStateCreateFlagsNV is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

PipelineCoverageToColorStateCreateInfoNV

Constructors

PipelineCoverageToColorStateCreateFlagsNV Flags

type NV_FRAGMENT_COVERAGE_TO_COLOR_SPEC_VERSION = 1

pattern NV_FRAGMENT_COVERAGE_TO_COLOR_SPEC_VERSION :: forall a. Integral a => a

type NV_FRAGMENT_COVERAGE_TO_COLOR_EXTENSION_NAME = "VK_NV_fragment_coverage_to_color"

pattern NV_FRAGMENT_COVERAGE_TO_COLOR_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a