Graphics.Vulkan.Extensions.VK_NV_shading_rate_image

Documentation

cmdBindShadingRateImageNV :: CommandBuffer -> ImageView -> ImageLayout -> IO ()

vkCmdBindShadingRateImageNV - Bind a shading rate image on a command buffer

Parameters

CommandBuffer is the command buffer into which the command will be recorded.

ImageView is an image view handle specifying the shading rate image. ImageView may be set to NULL_HANDLE, which is equivalent to specifying a view of an image filled with zero values.

ImageLayout is the layout that the image subresources accessible from ImageView will be in when the shading rate image is accessed.

Valid Usage

The shading rate image feature must be enabled.

If ImageView is not NULL_HANDLE, it must be a valid ImageView handle of type IMAGE_VIEW_TYPE_2D or IMAGE_VIEW_TYPE_2D_ARRAY.

If ImageView is not NULL_HANDLE, it must have a format of FORMAT_R8_UINT.

If ImageView is not NULL_HANDLE, it must have been created with a usage value including IMAGE_USAGE_SHADING_RATE_IMAGE_BIT_NV

If ImageView is not NULL_HANDLE, ImageLayout must match the actual ImageLayout of each subresource accessible from ImageView at the time the subresource is accessed.

If ImageView is not NULL_HANDLE, ImageLayout must be IMAGE_LAYOUT_SHADING_RATE_OPTIMAL_NV or IMAGE_LAYOUT_GENERAL.

Valid Usage (Implicit)

CommandBuffer must be a valid CommandBuffer handle

If ImageView is not NULL_HANDLE, ImageView must be a valid ImageView handle

ImageLayout must be a valid ImageLayout value

CommandBuffer must be in the recording state

The CommandPool that CommandBuffer was allocated from must support graphics operations

Both of CommandBuffer, and ImageView that are valid handles of non-ignored parameters must have been created, allocated, or retrieved from the same Device

Host Synchronization

Host access to CommandBuffer must be externally synchronized

Host access to the CommandPool that CommandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

CommandBuffer, ImageLayout, ImageView

cmdSetViewportShadingRatePaletteNV :: CommandBuffer -> ("firstViewport" ::: Word32) -> ("shadingRatePalettes" ::: Vector ShadingRatePaletteNV) -> IO ()

vkCmdSetViewportShadingRatePaletteNV - Set shading rate image palettes on a command buffer

Parameters

CommandBuffer is the command buffer into which the command will be recorded.

firstViewport is the index of the first viewport whose shading rate palette is updated by the command.

viewportCount is the number of viewports whose shading rate palettes are updated by the command.

pShadingRatePalettes is a pointer to an array of ShadingRatePaletteNV structures defining the palette for each viewport.

Valid Usage

The shading rate image feature must be enabled.

firstViewport must be less than PhysicalDeviceLimits::maxViewports

The sum of firstViewport and viewportCount must be between 1 and PhysicalDeviceLimits::maxViewports, inclusive

If the multiple viewports feature is not enabled, firstViewport must be 0

If the multiple viewports feature is not enabled, viewportCount must be 1

Valid Usage (Implicit)

CommandBuffer must be a valid CommandBuffer handle

pShadingRatePalettes must be a valid pointer to an array of viewportCount valid ShadingRatePaletteNV structures

CommandBuffer must be in the recording state

The CommandPool that CommandBuffer was allocated from must support graphics operations

viewportCount must be greater than 0

Host Synchronization

Host access to CommandBuffer must be externally synchronized

Host access to the CommandPool that CommandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

CommandBuffer, ShadingRatePaletteNV

cmdSetCoarseSampleOrderNV :: CommandBuffer -> CoarseSampleOrderTypeNV -> ("customSampleOrders" ::: Vector CoarseSampleOrderCustomNV) -> IO ()

vkCmdSetCoarseSampleOrderNV - Set sample order for coarse fragments on a command buffer

Parameters

CommandBuffer is the command buffer into which the command will be recorded.

sampleOrderType specifies the mechanism used to order coverage samples in fragments larger than one pixel.

customSampleOrderCount specifies the number of custom sample orderings to use when ordering coverage samples.

pCustomSampleOrders is a pointer to an array of CoarseSampleOrderCustomNV structures, each of which specifies the coverage sample order for a single combination of fragment area and coverage sample count.

Description

If sampleOrderType is COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV, the coverage sample order used for any combination of fragment area and coverage sample count not enumerated in pCustomSampleOrders will be identical to that used for COARSE_SAMPLE_ORDER_TYPE_DEFAULT_NV.

Valid Usage

If sampleOrderType is not COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV, customSamplerOrderCount must be 0

The array pCustomSampleOrders must not contain two structures with matching values for both the shadingRate and sampleCount members.

Valid Usage (Implicit)

CommandBuffer must be a valid CommandBuffer handle

sampleOrderType must be a valid CoarseSampleOrderTypeNV value

If customSampleOrderCount is not 0, pCustomSampleOrders must be a valid pointer to an array of customSampleOrderCount valid CoarseSampleOrderCustomNV structures

CommandBuffer must be in the recording state

The CommandPool that CommandBuffer was allocated from must support graphics operations

Host Synchronization

Host access to CommandBuffer must be externally synchronized

Host access to the CommandPool that CommandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

CoarseSampleOrderCustomNV, CoarseSampleOrderTypeNV, CommandBuffer

data ShadingRatePaletteNV

VkShadingRatePaletteNV - Structure specifying a single shading rate palette

Valid Usage (Implicit)

See Also

PipelineViewportShadingRateImageStateCreateInfoNV, ShadingRatePaletteEntryNV, cmdSetViewportShadingRatePaletteNV

Constructors

ShadingRatePaletteNV

Fields

shadingRatePaletteEntries :: Vector ShadingRatePaletteEntryNV

pShadingRatePaletteEntries must be a valid pointer to an array of shadingRatePaletteEntryCount valid ShadingRatePaletteEntryNV values

data PipelineViewportShadingRateImageStateCreateInfoNV

VkPipelineViewportShadingRateImageStateCreateInfoNV - Structure specifying parameters controlling shading rate image usage

Description

If this structure is not present, shadingRateImageEnable is considered to be FALSE, and the shading rate image and palettes are not used.

Valid Usage

If the multiple viewports feature is not enabled, viewportCount must be 0 or 1

viewportCount must be less than or equal to PhysicalDeviceLimits::maxViewports

If shadingRateImageEnable is TRUE, viewportCount must be equal to the viewportCount member of PipelineViewportStateCreateInfo

If no element of the pDynamicStates member of pDynamicState is DYNAMIC_STATE_VIEWPORT_SHADING_RATE_PALETTE_NV, pShadingRatePalettes must be a valid pointer to an array of viewportCount ShadingRatePaletteNV structures

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PIPELINE_VIEWPORT_SHADING_RATE_IMAGE_STATE_CREATE_INFO_NV

If viewportCount is not 0, and pShadingRatePalettes is not NULL, pShadingRatePalettes must be a valid pointer to an array of viewportCount valid ShadingRatePaletteNV structures

See Also

Bool32, ShadingRatePaletteNV, StructureType

Constructors

PipelineViewportShadingRateImageStateCreateInfoNV

Fields

shadingRateImageEnable :: Bool

shadingRateImageEnable specifies whether shading rate image and palettes are used during rasterization.

shadingRatePalettes :: Either Word32 (Vector ShadingRatePaletteNV)

pShadingRatePalettes is a pointer to an array of ShadingRatePaletteNV structures defining the palette for each viewport. If the shading rate palette state is dynamic, this member is ignored.

data PhysicalDeviceShadingRateImageFeaturesNV

VkPhysicalDeviceShadingRateImageFeaturesNV - Structure describing shading rate image features that can be supported by an implementation

Members

The members of the PhysicalDeviceShadingRateImageFeaturesNV structure describe the following features:

Description

See Shading Rate Image for more information.

If the PhysicalDeviceShadingRateImageFeaturesNV structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether the feature is supported. PhysicalDeviceShadingRateImageFeaturesNV can also be included in the pNext chain of DeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceShadingRateImageFeaturesNV

Fields

shadingRateImage :: Bool

shadingRateImage indicates that the implementation supports the use of a shading rate image to derive an effective shading rate for fragment processing. It also indicates that the implementation supports the ShadingRateNV SPIR-V execution mode.

shadingRateCoarseSampleOrder :: Bool

shadingRateCoarseSampleOrder indicates that the implementation supports a user-configurable ordering of coverage samples in fragments larger than one pixel.

data PhysicalDeviceShadingRateImagePropertiesNV

VkPhysicalDeviceShadingRateImagePropertiesNV - Structure describing shading rate image limits that can be supported by an implementation

Members

The members of the PhysicalDeviceShadingRateImagePropertiesNV structure describe the following implementation-dependent properties related to the shading rate image feature:

Description

If the PhysicalDeviceShadingRateImagePropertiesNV structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

Extent2D, StructureType

Constructors

PhysicalDeviceShadingRateImagePropertiesNV

Fields

shadingRateTexelSize :: Extent2D

shadingRateTexelSize indicates the width and height of the portion of the framebuffer corresponding to each texel in the shading rate image.

shadingRatePaletteSize :: Word32

shadingRatePaletteSize indicates the maximum number of palette entries supported for the shading rate image.

shadingRateMaxCoarseSamples :: Word32

shadingRateMaxCoarseSamples specifies the maximum number of coverage samples supported in a single fragment. If the product of the fragment size derived from the base shading rate and the number of coverage samples per pixel exceeds this limit, the final shading rate will be adjusted so that its product does not exceed the limit.

data CoarseSampleLocationNV

VkCoarseSampleLocationNV - Structure specifying parameters controlling shading rate image usage

Valid Usage

See Also

CoarseSampleOrderCustomNV

Constructors

CoarseSampleLocationNV

Fields

pixelX :: Word32

pixelX must be less than the width (in pixels) of the fragment.

pixelY :: Word32

pixelY must be less than the height (in pixels) of the fragment.

sample :: Word32

sample must be less than the number of coverage samples in each pixel belonging to the fragment.

data CoarseSampleOrderCustomNV

VkCoarseSampleOrderCustomNV - Structure specifying parameters controlling shading rate image usage

Description

When using a custom sample ordering, element i in pSampleLocations specifies a specific pixel and per-pixel coverage sample number that corresponds to the coverage sample numbered i in the multi-pixel fragment.

Valid Usage

shadingRate must be a shading rate that generates fragments with more than one pixel.

sampleCount must correspond to a sample count enumerated in SampleCountFlags whose corresponding bit is set in PhysicalDeviceLimits::framebufferNoAttachmentsSampleCounts.

sampleLocationCount must be equal to the product of sampleCount, the fragment width for shadingRate, and the fragment height for shadingRate.

sampleLocationCount must be less than or equal to the value of PhysicalDeviceShadingRateImagePropertiesNV::shadingRateMaxCoarseSamples.

The array pSampleLocations must contain exactly one entry for every combination of valid values for pixelX, pixelY, and sample in the structure CoarseSampleOrderCustomNV.

Valid Usage (Implicit)

shadingRate must be a valid ShadingRatePaletteEntryNV value

pSampleLocations must be a valid pointer to an array of sampleLocationCount CoarseSampleLocationNV structures

sampleLocationCount must be greater than 0

See Also

CoarseSampleLocationNV, PipelineViewportCoarseSampleOrderStateCreateInfoNV, ShadingRatePaletteEntryNV, cmdSetCoarseSampleOrderNV

Constructors

CoarseSampleOrderCustomNV

Fields

shadingRate :: ShadingRatePaletteEntryNV

shadingRate is a shading rate palette entry that identifies the fragment width and height for the combination of fragment area and per-pixel coverage sample count to control.

sampleCount :: Word32

sampleCount identifies the per-pixel coverage sample count for the combination of fragment area and coverage sample count to control.

sampleLocations :: Vector CoarseSampleLocationNV

pSampleLocations is a pointer to an array of CoarseSampleOrderCustomNV structures specifying the location of each sample in the custom ordering.

data PipelineViewportCoarseSampleOrderStateCreateInfoNV

VkPipelineViewportCoarseSampleOrderStateCreateInfoNV - Structure specifying parameters controlling sample order in coarse fragments

Description

If this structure is not present, sampleOrderType is considered to be COARSE_SAMPLE_ORDER_TYPE_DEFAULT_NV.

If sampleOrderType is COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV, the coverage sample order used for any combination of fragment area and coverage sample count not enumerated in pCustomSampleOrders will be identical to that used for COARSE_SAMPLE_ORDER_TYPE_DEFAULT_NV.

If the pipeline was created with DYNAMIC_STATE_VIEWPORT_COARSE_SAMPLE_ORDER_NV, the contents of this structure (if present) are ignored, and the coverage sample order is instead specified by cmdSetCoarseSampleOrderNV.

Valid Usage

If sampleOrderType is not COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV, customSamplerOrderCount must be 0

The array pCustomSampleOrders must not contain two structures with matching values for both the shadingRate and sampleCount members.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PIPELINE_VIEWPORT_COARSE_SAMPLE_ORDER_STATE_CREATE_INFO_NV

sampleOrderType must be a valid CoarseSampleOrderTypeNV value

If customSampleOrderCount is not 0, pCustomSampleOrders must be a valid pointer to an array of customSampleOrderCount valid CoarseSampleOrderCustomNV structures

See Also

CoarseSampleOrderCustomNV, CoarseSampleOrderTypeNV, StructureType

Constructors

PipelineViewportCoarseSampleOrderStateCreateInfoNV

Fields

sampleOrderType :: CoarseSampleOrderTypeNV

sampleOrderType specifies the mechanism used to order coverage samples in fragments larger than one pixel.

customSampleOrders :: Vector CoarseSampleOrderCustomNV

pCustomSampleOrders is a pointer to an array of customSampleOrderCount CoarseSampleOrderCustomNV structures, each of which specifies the coverage sample order for a single combination of fragment area and coverage sample count.

newtype ShadingRatePaletteEntryNV

VkShadingRatePaletteEntryNV - Shading rate image palette entry types

Description

The following table indicates the width and height (in pixels) of each fragment generated using the indicated shading rate, as well as the maximum number of fragment shader invocations launched for each fragment. When processing regions of a primitive that have a shading rate of SHADING_RATE_PALETTE_ENTRY_NO_INVOCATIONS_NV, no fragments will be generated in that region.

Shading Rate 	 Width 	 Height 	 Invocations

SHADING_RATE_PALETTE_ENTRY_NO_INVOCATIONS_NV 	 0 	 0 	 0

SHADING_RATE_PALETTE_ENTRY_16_INVOCATIONS_PER_PIXEL_NV 	 1 	 1 	 16

SHADING_RATE_PALETTE_ENTRY_8_INVOCATIONS_PER_PIXEL_NV 	 1 	 1 	 8

SHADING_RATE_PALETTE_ENTRY_4_INVOCATIONS_PER_PIXEL_NV 	 1 	 1 	 4

SHADING_RATE_PALETTE_ENTRY_2_INVOCATIONS_PER_PIXEL_NV 	 1 	 1 	 2

SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_PIXEL_NV 	 1 	 1 	 1

SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_2X1_PIXELS_NV 	 2 	 1 	 1

SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_1X2_PIXELS_NV 	 1 	 2 	 1

SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_2X2_PIXELS_NV 	 2 	 2 	 1

SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_4X2_PIXELS_NV 	 4 	 2 	 1

SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_2X4_PIXELS_NV 	 2 	 4 	 1

SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_4X4_PIXELS_NV 	 4 	 4 	 1

See Also

CoarseSampleOrderCustomNV, ShadingRatePaletteNV

Constructors

ShadingRatePaletteEntryNV Int32

Bundled Patterns

pattern SHADING_RATE_PALETTE_ENTRY_NO_INVOCATIONS_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_16_INVOCATIONS_PER_PIXEL_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_8_INVOCATIONS_PER_PIXEL_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_4_INVOCATIONS_PER_PIXEL_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_2_INVOCATIONS_PER_PIXEL_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_PIXEL_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_2X1_PIXELS_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_1X2_PIXELS_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_2X2_PIXELS_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_4X2_PIXELS_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_2X4_PIXELS_NV :: ShadingRatePaletteEntryNV

pattern SHADING_RATE_PALETTE_ENTRY_1_INVOCATION_PER_4X4_PIXELS_NV :: ShadingRatePaletteEntryNV

newtype CoarseSampleOrderTypeNV

VkCoarseSampleOrderTypeNV - Shading rate image sample ordering types

See Also

PipelineViewportCoarseSampleOrderStateCreateInfoNV, cmdSetCoarseSampleOrderNV

Constructors

CoarseSampleOrderTypeNV Int32

Bundled Patterns

pattern COARSE_SAMPLE_ORDER_TYPE_DEFAULT_NV :: CoarseSampleOrderTypeNV
COARSE_SAMPLE_ORDER_TYPE_DEFAULT_NV specifies that coverage samples will be ordered in an implementation-dependent manner.

pattern COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV :: CoarseSampleOrderTypeNV
COARSE_SAMPLE_ORDER_TYPE_CUSTOM_NV specifies that coverage samples will be ordered according to the array of custom orderings provided in either the pCustomSampleOrders member of PipelineViewportCoarseSampleOrderStateCreateInfoNV or the pCustomSampleOrders member of cmdSetCoarseSampleOrderNV.

pattern COARSE_SAMPLE_ORDER_TYPE_PIXEL_MAJOR_NV :: CoarseSampleOrderTypeNV
COARSE_SAMPLE_ORDER_TYPE_PIXEL_MAJOR_NV specifies that coverage samples will be ordered sequentially, sorted first by pixel coordinate (in row-major order) and then by coverage sample number.

pattern COARSE_SAMPLE_ORDER_TYPE_SAMPLE_MAJOR_NV :: CoarseSampleOrderTypeNV
COARSE_SAMPLE_ORDER_TYPE_SAMPLE_MAJOR_NV specifies that coverage samples will be ordered sequentially, sorted first by coverage sample number and then by pixel coordinate (in row-major order).

type NV_SHADING_RATE_IMAGE_SPEC_VERSION = 3

pattern NV_SHADING_RATE_IMAGE_SPEC_VERSION :: forall a. Integral a => a

type NV_SHADING_RATE_IMAGE_EXTENSION_NAME = "VK_NV_shading_rate_image"

pattern NV_SHADING_RATE_IMAGE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a