Vulkan.Core10.BufferView

Documentation

createBufferView :: forall io. MonadIO io => Device -> BufferViewCreateInfo -> ("allocator" ::: Maybe AllocationCallbacks) -> io BufferView

vkCreateBufferView - Create a new buffer view object

Parameters

device is the logical device that creates the buffer view.

pCreateInfo is a pointer to a BufferViewCreateInfo structure containing parameters to be used to create the buffer.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

pView is a pointer to a BufferView handle in which the resulting buffer view object is returned.

Valid Usage (Implicit)

device must be a valid Device handle

pCreateInfo must be a valid pointer to a valid BufferViewCreateInfo structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pView must be a valid pointer to a BufferView handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, BufferView, BufferViewCreateInfo, Device

withBufferView :: forall io r. MonadIO io => Device -> BufferViewCreateInfo -> Maybe AllocationCallbacks -> (io BufferView -> (BufferView -> io ()) -> r) -> r

A convenience wrapper to make a compatible pair of calls to createBufferView and destroyBufferView

To ensure that destroyBufferView is always called: pass bracket (or the allocate function from your favourite resource management library) as the first argument. To just extract the pair pass (,) as the first argument.

destroyBufferView :: forall io. MonadIO io => Device -> BufferView -> ("allocator" ::: Maybe AllocationCallbacks) -> io ()

vkDestroyBufferView - Destroy a buffer view object

Parameters

device is the logical device that destroys the buffer view.

bufferView is the buffer view to destroy.

pAllocator controls host memory allocation as described in the Memory Allocation chapter.

Valid Usage

All submitted commands that refer to bufferView must have completed execution

If AllocationCallbacks were provided when bufferView was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when bufferView was created, pAllocator must be NULL

Valid Usage (Implicit)

device must be a valid Device handle

If bufferView is not NULL_HANDLE, bufferView must be a valid BufferView handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If bufferView is a valid handle, it must have been created, allocated, or retrieved from device

Host Synchronization

Host access to bufferView must be externally synchronized

See Also

AllocationCallbacks, BufferView, Device

data BufferViewCreateInfo

VkBufferViewCreateInfo - Structure specifying parameters of a newly created buffer view

Valid Usage

offset must be less than the size of buffer

If range is not equal to WHOLE_SIZE, range must be greater than 0

If range is not equal to WHOLE_SIZE, range must be an integer multiple of the texel block size of format

If range is not equal to WHOLE_SIZE, range divided by the texel block size of format, multiplied by the number of texels per texel block for that format (as defined in the Compatible Formats table), must be less than or equal to PhysicalDeviceLimits::maxTexelBufferElements

If range is not equal to WHOLE_SIZE, the sum of offset and range must be less than or equal to the size of buffer

buffer must have been created with a usage value containing at least one of BUFFER_USAGE_UNIFORM_TEXEL_BUFFER_BIT or BUFFER_USAGE_STORAGE_TEXEL_BUFFER_BIT

If buffer was created with usage containing BUFFER_USAGE_UNIFORM_TEXEL_BUFFER_BIT, format must be supported for uniform texel buffers, as specified by the FORMAT_FEATURE_UNIFORM_TEXEL_BUFFER_BIT flag in FormatProperties::bufferFeatures returned by getPhysicalDeviceFormatProperties

If buffer was created with usage containing BUFFER_USAGE_STORAGE_TEXEL_BUFFER_BIT, format must be supported for storage texel buffers, as specified by the FORMAT_FEATURE_STORAGE_TEXEL_BUFFER_BIT flag in FormatProperties::bufferFeatures returned by getPhysicalDeviceFormatProperties

If buffer is non-sparse then it must be bound completely and contiguously to a single DeviceMemory object

If the texelBufferAlignment feature is not enabled, offset must be a multiple of PhysicalDeviceLimits::minTexelBufferOffsetAlignment

If the texelBufferAlignment feature is enabled and if buffer was created with usage containing BUFFER_USAGE_STORAGE_TEXEL_BUFFER_BIT, offset must be a multiple of the lesser of PhysicalDeviceTexelBufferAlignmentPropertiesEXT::storageTexelBufferOffsetAlignmentBytes or, if PhysicalDeviceTexelBufferAlignmentPropertiesEXT::storageTexelBufferOffsetSingleTexelAlignment is TRUE, the size of a texel of the requested format. If the size of a texel is a multiple of three bytes, then the size of a single component of format is used instead

If the texelBufferAlignment feature is enabled and if buffer was created with usage containing BUFFER_USAGE_UNIFORM_TEXEL_BUFFER_BIT, offset must be a multiple of the lesser of PhysicalDeviceTexelBufferAlignmentPropertiesEXT::uniformTexelBufferOffsetAlignmentBytes or, if PhysicalDeviceTexelBufferAlignmentPropertiesEXT::uniformTexelBufferOffsetSingleTexelAlignment is TRUE, the size of a texel of the requested format. If the size of a texel is a multiple of three bytes, then the size of a single component of format is used instead

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_BUFFER_VIEW_CREATE_INFO

pNext must be NULL

flags must be 0

buffer must be a valid Buffer handle

format must be a valid Format value

See Also

Buffer, BufferViewCreateFlags, DeviceSize, Format, StructureType, createBufferView

Constructors

BufferViewCreateInfo

Fields

flags :: BufferViewCreateFlags

flags is reserved for future use.

buffer :: Buffer

buffer is a Buffer on which the view will be created.

format :: Format

format is a Format describing the format of the data elements in the buffer.

offset :: DeviceSize

offset is an offset in bytes from the base address of the buffer. Accesses to the buffer view from shaders use addressing that is relative to this starting offset.

range :: DeviceSize

range is a size in bytes of the buffer view. If range is equal to WHOLE_SIZE, the range from offset to the end of the buffer is used. If WHOLE_SIZE is used and the remaining size of the buffer is not a multiple of the texel block size of format, the nearest smaller multiple is used.