Vulkan.Extensions.VK_AMD_rasterization_order

Documentation

data PipelineRasterizationStateRasterizationOrderAMD

VkPipelineRasterizationStateRasterizationOrderAMD - Structure defining rasterization order for a graphics pipeline

Valid Usage (Implicit)

If the VK_AMD_rasterization_order device extension is not enabled or the application does not request a particular rasterization order through specifying a PipelineRasterizationStateRasterizationOrderAMD structure then the rasterization order used by the graphics pipeline defaults to RASTERIZATION_ORDER_STRICT_AMD.

See Also

RasterizationOrderAMD, StructureType

Constructors

PipelineRasterizationStateRasterizationOrderAMD

Fields

rasterizationOrder :: RasterizationOrderAMD

rasterizationOrder is a RasterizationOrderAMD value specifying the primitive rasterization order to use.

rasterizationOrder must be a valid RasterizationOrderAMD value

newtype RasterizationOrderAMD

VkRasterizationOrderAMD - Specify rasterization order for a graphics pipeline

See Also

PipelineRasterizationStateRasterizationOrderAMD

Constructors

RasterizationOrderAMD Int32

Bundled Patterns

pattern RASTERIZATION_ORDER_STRICT_AMD :: RasterizationOrderAMD
RASTERIZATION_ORDER_STRICT_AMD specifies that operations for each primitive in a subpass must occur in primitive order.

pattern RASTERIZATION_ORDER_RELAXED_AMD :: RasterizationOrderAMD
RASTERIZATION_ORDER_RELAXED_AMD specifies that operations for each primitive in a subpass may not occur in primitive order.

type AMD_RASTERIZATION_ORDER_SPEC_VERSION = 1

pattern AMD_RASTERIZATION_ORDER_SPEC_VERSION :: forall a. Integral a => a

type AMD_RASTERIZATION_ORDER_EXTENSION_NAME = "VK_AMD_rasterization_order"

pattern AMD_RASTERIZATION_ORDER_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a