Vulkan.Extensions.VK_AMD_shader_core_properties

Documentation

data PhysicalDeviceShaderCorePropertiesAMD

VkPhysicalDeviceShaderCorePropertiesAMD - Structure describing shader core properties that can be supported by an implementation

Members

The members of the PhysicalDeviceShaderCorePropertiesAMD structure describe the following implementation-dependent limits:

Description

If the PhysicalDeviceShaderCorePropertiesAMD structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

StructureType

Constructors

PhysicalDeviceShaderCorePropertiesAMD

Fields

shaderEngineCount :: Word32

shaderEngineCount is an unsigned integer value indicating the number of shader engines found inside the shader core of the physical device.

shaderArraysPerEngineCount :: Word32

shaderArraysPerEngineCount is an unsigned integer value indicating the number of shader arrays inside a shader engine. Each shader array has its own scan converter, set of compute units, and a render back end (color and depth buffers). Shader arrays within a shader engine share shader processor input (wave launcher) and shader export (export buffer) units. Currently, a shader engine can have one or two shader arrays.

computeUnitsPerShaderArray :: Word32

computeUnitsPerShaderArray is an unsigned integer value indicating the physical number of compute units within a shader array. The active number of compute units in a shader array may be lower. A compute unit houses a set of SIMDs along with a sequencer module and a local data store.

simdPerComputeUnit :: Word32

simdPerComputeUnit is an unsigned integer value indicating the number of SIMDs inside a compute unit. Each SIMD processes a single instruction at a time.

wavefrontsPerSimd :: Word32

wavefrontSize :: Word32

wavefrontSize is an unsigned integer value indicating the maximum size of a subgroup.

sgprsPerSimd :: Word32

sgprsPerSimd is an unsigned integer value indicating the number of physical Scalar General Purpose Registers (SGPRs) per SIMD.

minSgprAllocation :: Word32

minSgprAllocation is an unsigned integer value indicating the minimum number of SGPRs allocated for a wave.

maxSgprAllocation :: Word32

maxSgprAllocation is an unsigned integer value indicating the maximum number of SGPRs allocated for a wave.

sgprAllocationGranularity :: Word32

sgprAllocationGranularity is an unsigned integer value indicating the granularity of SGPR allocation for a wave.

vgprsPerSimd :: Word32

vgprsPerSimd is an unsigned integer value indicating the number of physical Vector General Purpose Registers (VGPRs) per SIMD.

minVgprAllocation :: Word32

minVgprAllocation is an unsigned integer value indicating the minimum number of VGPRs allocated for a wave.

maxVgprAllocation :: Word32

maxVgprAllocation is an unsigned integer value indicating the maximum number of VGPRs allocated for a wave.

vgprAllocationGranularity :: Word32

vgprAllocationGranularity is an unsigned integer value indicating the granularity of VGPR allocation for a wave.

type AMD_SHADER_CORE_PROPERTIES_SPEC_VERSION = 2

pattern AMD_SHADER_CORE_PROPERTIES_SPEC_VERSION :: forall a. Integral a => a

type AMD_SHADER_CORE_PROPERTIES_EXTENSION_NAME = "VK_AMD_shader_core_properties"

pattern AMD_SHADER_CORE_PROPERTIES_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a