Vulkan.Extensions.VK_EXT_buffer_device_address

Documentation

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_BUFFER_ADDRESS_FEATURES_EXT :: StructureType

pattern STRUCTURE_TYPE_BUFFER_DEVICE_ADDRESS_INFO_EXT :: StructureType

pattern BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT_EXT :: BufferUsageFlagBits

pattern BUFFER_CREATE_DEVICE_ADDRESS_CAPTURE_REPLAY_BIT_EXT :: BufferCreateFlagBits

pattern ERROR_INVALID_DEVICE_ADDRESS_EXT :: Result

getBufferDeviceAddressEXT :: MonadIO io => Device -> BufferDeviceAddressInfo -> io DeviceAddress

data PhysicalDeviceBufferDeviceAddressFeaturesEXT

VkPhysicalDeviceBufferDeviceAddressFeaturesEXT - Structure describing buffer address features that can be supported by an implementation

Members

The members of the PhysicalDeviceBufferDeviceAddressFeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceBufferDeviceAddressFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether the feature is supported. PhysicalDeviceBufferDeviceAddressFeaturesEXT can also be included in the pNext chain of DeviceCreateInfo to enable features.

Note

The PhysicalDeviceBufferDeviceAddressFeaturesEXT structure has the same members as the PhysicalDeviceBufferDeviceAddressFeatures structure, but the functionality indicated by the members is expressed differently. The features indicated by the PhysicalDeviceBufferDeviceAddressFeatures structure requires additional flags to be passed at memory allocation time, and the capture and replay mechanism is built around opaque capture addresses for buffer and memory objects.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceBufferDeviceAddressFeaturesEXT

Fields

bufferDeviceAddress :: Bool

bufferDeviceAddress indicates that the implementation supports accessing buffer memory in shaders as storage buffers via an address queried from getBufferDeviceAddressEXT.

bufferDeviceAddressCaptureReplay :: Bool

bufferDeviceAddressCaptureReplay indicates that the implementation supports saving and reusing buffer addresses, e.g. for trace capture and replay.

bufferDeviceAddressMultiDevice :: Bool

bufferDeviceAddressMultiDevice indicates that the implementation supports the bufferDeviceAddress feature for logical devices created with multiple physical devices. If this feature is not supported, buffer addresses must not be queried on a logical device created with more than one physical device.

data BufferDeviceAddressCreateInfoEXT

VkBufferDeviceAddressCreateInfoEXT - Request a specific address for a buffer

Description

If deviceAddress is zero, no specific address is requested.

If deviceAddress is not zero, then it must be an address retrieved from an identically created buffer on the same implementation. The buffer must also be bound to an identically created DeviceMemory object.

If this structure is not present, it is as if deviceAddress is zero.

Apps should avoid creating buffers with app-provided addresses and implementation-provided addresses in the same process, to reduce the likelihood of ERROR_INVALID_DEVICE_ADDRESS_EXT errors.

Valid Usage (Implicit)

See Also

DeviceAddress, StructureType

Constructors

BufferDeviceAddressCreateInfoEXT

Fields

deviceAddress :: DeviceAddress

deviceAddress is the device address requested for the buffer.

type PhysicalDeviceBufferAddressFeaturesEXT = PhysicalDeviceBufferDeviceAddressFeaturesEXT

type BufferDeviceAddressInfoEXT = BufferDeviceAddressInfo

type EXT_BUFFER_DEVICE_ADDRESS_SPEC_VERSION = 2

pattern EXT_BUFFER_DEVICE_ADDRESS_SPEC_VERSION :: forall a. Integral a => a

type EXT_BUFFER_DEVICE_ADDRESS_EXTENSION_NAME = "VK_EXT_buffer_device_address"

pattern EXT_BUFFER_DEVICE_ADDRESS_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a