Vulkan.Extensions.VK_EXT_custom_border_color

Documentation

data SamplerCustomBorderColorCreateInfoEXT

VkSamplerCustomBorderColorCreateInfoEXT - Structure specifying custom border color

Valid Usage

If provided format is not FORMAT_UNDEFINED then the SamplerCreateInfo::borderColor type must match the sampled type of the provided format, as shown in the SPIR-V Sampled Type column of the https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#formats-numericformat table

If the customBorderColorWithoutFormat feature is not enabled then format must not be FORMAT_UNDEFINED

If the sampler is used to sample an image view of FORMAT_B4G4R4A4_UNORM_PACK16, FORMAT_B5G6R5_UNORM_PACK16, or FORMAT_B5G5R5A1_UNORM_PACK16 format then format must not be FORMAT_UNDEFINED

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_SAMPLER_CUSTOM_BORDER_COLOR_CREATE_INFO_EXT

format must be a valid Format value

See Also

ClearColorValue, Format, StructureType

Constructors

SamplerCustomBorderColorCreateInfoEXT

Fields

customBorderColor :: ClearColorValue

customBorderColor is a ClearColorValue representing the desired custom sampler border color.

format :: Format

format is a Format representing the format of the sampled image view(s). This field may be FORMAT_UNDEFINED if the customBorderColorWithoutFormat feature is enabled.

data PhysicalDeviceCustomBorderColorPropertiesEXT

VkPhysicalDeviceCustomBorderColorPropertiesEXT - Structure describing whether custom border colors can be supported by an implementation

Members

The members of the PhysicalDeviceCustomBorderColorPropertiesEXT structure describe the following features:

Valid Usage (Implicit)

See Also

StructureType

Constructors

PhysicalDeviceCustomBorderColorPropertiesEXT

Fields

maxCustomBorderColorSamplers :: Word32

maxCustomBorderColorSamplers indicates the maximum number of samplers with custom border colors which can simultaneously exist on a device.

data PhysicalDeviceCustomBorderColorFeaturesEXT

VkPhysicalDeviceCustomBorderColorFeaturesEXT - Structure describing whether custom border colors can be supported by an implementation

Members

The members of the PhysicalDeviceCustomBorderColorFeaturesEXT structure describe the following features:

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceCustomBorderColorFeaturesEXT

Fields

customBorderColors :: Bool

customBorderColors indicates that the implementation supports providing a borderColor value with one of the following values at sampler creation time:

BORDER_COLOR_FLOAT_CUSTOM_EXT

BORDER_COLOR_INT_CUSTOM_EXT

customBorderColorWithoutFormat :: Bool

customBorderColorWithoutFormat indicates that explicit formats are not required for custom border colors and the value of the format member of the SamplerCustomBorderColorCreateInfoEXT structure may be FORMAT_UNDEFINED. If this feature bit is not set, applications must provide the Format of the image view(s) being sampled by this sampler in the format member of the SamplerCustomBorderColorCreateInfoEXT structure.

type EXT_CUSTOM_BORDER_COLOR_SPEC_VERSION = 12

pattern EXT_CUSTOM_BORDER_COLOR_SPEC_VERSION :: forall a. Integral a => a

type EXT_CUSTOM_BORDER_COLOR_EXTENSION_NAME = "VK_EXT_custom_border_color"

pattern EXT_CUSTOM_BORDER_COLOR_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a