Vulkan.Extensions.VK_EXT_debug_utils

Documentation

setDebugUtilsObjectNameEXT

Arguments

:: forall io. MonadIO io

=> Device
device is the device that created the object.

-> DebugUtilsObjectNameInfoEXT
pNameInfo is a pointer to a DebugUtilsObjectNameInfoEXT structure specifying parameters of the name to set on the object.

-> io ()

vkSetDebugUtilsObjectNameEXT - Give a user-friendly name to an object

Valid Usage

pNameInfo->objectType must not be OBJECT_TYPE_UNKNOWN

pNameInfo->objectHandle must not be NULL_HANDLE

Valid Usage (Implicit)

device must be a valid Device handle

pNameInfo must be a valid pointer to a valid DebugUtilsObjectNameInfoEXT structure

Host Synchronization

Host access to pNameInfo->objectHandle must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

DebugUtilsObjectNameInfoEXT, Device

setDebugUtilsObjectTagEXT

Arguments

:: forall io. MonadIO io

=> Device
device is the device that created the object.

-> DebugUtilsObjectTagInfoEXT
pTagInfo is a pointer to a DebugUtilsObjectTagInfoEXT structure specifying parameters of the tag to attach to the object.

-> io ()

vkSetDebugUtilsObjectTagEXT - Attach arbitrary data to an object

Valid Usage (Implicit)

device must be a valid Device handle

pTagInfo must be a valid pointer to a valid DebugUtilsObjectTagInfoEXT structure

Host Synchronization

Host access to pTagInfo->objectHandle must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

DebugUtilsObjectTagInfoEXT, Device

queueBeginDebugUtilsLabelEXT

Arguments

:: forall io. MonadIO io

=> Queue
queue is the queue in which to start a debug label region.

queue must be a valid Queue handle

-> ("labelInfo" ::: DebugUtilsLabelEXT)
pLabelInfo is a pointer to a DebugUtilsLabelEXT structure specifying parameters of the label region to open.

pLabelInfo must be a valid pointer to a valid DebugUtilsLabelEXT structure

-> io ()

vkQueueBeginDebugUtilsLabelEXT - Open a queue debug label region

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

- 	 - 	 Any 	 -

See Also

DebugUtilsLabelEXT, Queue

queueEndDebugUtilsLabelEXT

Arguments

:: forall io. MonadIO io

=> Queue
queue is the queue in which a debug label region should be closed.

-> io ()

vkQueueEndDebugUtilsLabelEXT - Close a queue debug label region

Description

The calls to queueBeginDebugUtilsLabelEXT and queueEndDebugUtilsLabelEXT must be matched and balanced.

Valid Usage

There must be an outstanding queueBeginDebugUtilsLabelEXT command prior to the queueEndDebugUtilsLabelEXT on the queue

Valid Usage (Implicit)

queue must be a valid Queue handle

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

- 	 - 	 Any 	 -

See Also

Queue

queueInsertDebugUtilsLabelEXT

Arguments

:: forall io. MonadIO io

=> Queue
queue is the queue into which a debug label will be inserted.

queue must be a valid Queue handle

-> ("labelInfo" ::: DebugUtilsLabelEXT)
pLabelInfo is a pointer to a DebugUtilsLabelEXT structure specifying parameters of the label to insert.

pLabelInfo must be a valid pointer to a valid DebugUtilsLabelEXT structure

-> io ()

vkQueueInsertDebugUtilsLabelEXT - Insert a label into a queue

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

- 	 - 	 Any 	 -

See Also

DebugUtilsLabelEXT, Queue

cmdBeginDebugUtilsLabelEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command is recorded.

-> ("labelInfo" ::: DebugUtilsLabelEXT)
pLabelInfo is a pointer to a DebugUtilsLabelEXT structure specifying parameters of the label region to open.

-> io ()

vkCmdBeginDebugUtilsLabelEXT - Open a command buffer debug label region

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pLabelInfo must be a valid pointer to a valid DebugUtilsLabelEXT structure

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute

See Also

CommandBuffer, DebugUtilsLabelEXT

cmdUseDebugUtilsLabelEXT :: forall io r. MonadIO io => CommandBuffer -> DebugUtilsLabelEXT -> io r -> io r

This function will call the supplied action between calls to cmdBeginDebugUtilsLabelEXT and cmdEndDebugUtilsLabelEXT

Note that cmdEndDebugUtilsLabelEXT is *not* called if an exception is thrown by the inner action.

cmdEndDebugUtilsLabelEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command is recorded.

-> io ()

vkCmdEndDebugUtilsLabelEXT - Close a command buffer label region

Description

An application may open a debug label region in one command buffer and close it in another, or otherwise split debug label regions across multiple command buffers or multiple queue submissions. When viewed from the linear series of submissions to a single queue, the calls to cmdBeginDebugUtilsLabelEXT and cmdEndDebugUtilsLabelEXT must be matched and balanced.

Valid Usage

There must be an outstanding cmdBeginDebugUtilsLabelEXT command prior to the cmdEndDebugUtilsLabelEXT on the queue that commandBuffer is submitted to

If commandBuffer is a secondary command buffer, there must be an outstanding cmdBeginDebugUtilsLabelEXT command recorded to commandBuffer that has not previously been ended by a call to cmdEndDebugUtilsLabelEXT

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute

See Also

CommandBuffer

cmdInsertDebugUtilsLabelEXT

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command is recorded.

-> ("labelInfo" ::: DebugUtilsLabelEXT)

-> io ()

vkCmdInsertDebugUtilsLabelEXT - Insert a label into a command buffer

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pLabelInfo must be a valid pointer to a valid DebugUtilsLabelEXT structure

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics, or compute operations

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics Compute

See Also

CommandBuffer, DebugUtilsLabelEXT

createDebugUtilsMessengerEXT

Arguments

:: forall io. MonadIO io

=> Instance
instance is the instance the messenger will be used with.

-> DebugUtilsMessengerCreateInfoEXT
pCreateInfo is a pointer to a DebugUtilsMessengerCreateInfoEXT structure containing the callback pointer, as well as defining conditions under which this messenger will trigger the callback.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io DebugUtilsMessengerEXT

vkCreateDebugUtilsMessengerEXT - Create a debug messenger object

Valid Usage (Implicit)

instance must be a valid Instance handle

pCreateInfo must be a valid pointer to a valid DebugUtilsMessengerCreateInfoEXT structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pMessenger must be a valid pointer to a DebugUtilsMessengerEXT handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

The application must ensure that createDebugUtilsMessengerEXT is not executed in parallel with any Vulkan command that is also called with instance or child of instance as the dispatchable argument.

See Also

AllocationCallbacks, DebugUtilsMessengerCreateInfoEXT, DebugUtilsMessengerEXT, Instance

withDebugUtilsMessengerEXT :: forall io r. MonadIO io => Instance -> DebugUtilsMessengerCreateInfoEXT -> Maybe AllocationCallbacks -> (io DebugUtilsMessengerEXT -> (DebugUtilsMessengerEXT -> io ()) -> r) -> r

A convenience wrapper to make a compatible pair of calls to createDebugUtilsMessengerEXT and destroyDebugUtilsMessengerEXT

To ensure that destroyDebugUtilsMessengerEXT is always called: pass bracket (or the allocate function from your favourite resource management library) as the first argument. To just extract the pair pass (,) as the first argument.

destroyDebugUtilsMessengerEXT

Arguments

:: forall io. MonadIO io

=> Instance
instance is the instance where the callback was created.

-> DebugUtilsMessengerEXT
messenger is the DebugUtilsMessengerEXT object to destroy. messenger is an externally synchronized object and must not be used on more than one thread at a time. This means that destroyDebugUtilsMessengerEXT must not be called when a callback is active.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator controls host memory allocation as described in the Memory Allocation chapter.

-> io ()

vkDestroyDebugUtilsMessengerEXT - Destroy a debug messenger object

Valid Usage

If AllocationCallbacks were provided when messenger was created, a compatible set of callbacks must be provided here

If no AllocationCallbacks were provided when messenger was created, pAllocator must be NULL

Valid Usage (Implicit)

instance must be a valid Instance handle

If messenger is not NULL_HANDLE, messenger must be a valid DebugUtilsMessengerEXT handle

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

If messenger is a valid handle, it must have been created, allocated, or retrieved from instance

Host Synchronization

Host access to messenger must be externally synchronized

The application must ensure that destroyDebugUtilsMessengerEXT is not executed in parallel with any Vulkan command that is also called with instance or child of instance as the dispatchable argument.

See Also

AllocationCallbacks, DebugUtilsMessengerEXT, Instance

submitDebugUtilsMessageEXT

Arguments

:: forall io. MonadIO io

=> Instance
instance is the debug stream’s Instance.

-> DebugUtilsMessageSeverityFlagBitsEXT
messageSeverity is the DebugUtilsMessageSeverityFlagBitsEXT severity of this event/message.

-> ("messageTypes" ::: DebugUtilsMessageTypeFlagsEXT)
messageTypes is a bitmask of DebugUtilsMessageTypeFlagBitsEXT specifying which type of event(s) to identify with this message.

-> DebugUtilsMessengerCallbackDataEXT
pCallbackData contains all the callback related data in the DebugUtilsMessengerCallbackDataEXT structure.

-> io ()

vkSubmitDebugUtilsMessageEXT - Inject a message into a debug stream

Description

The call will propagate through the layers and generate callback(s) as indicated by the message’s flags. The parameters are passed on to the callback in addition to the pUserData value that was defined at the time the messenger was registered.

Valid Usage

The objectType member of each element of pCallbackData->pObjects must not be OBJECT_TYPE_UNKNOWN

Valid Usage (Implicit)

instance must be a valid Instance handle

messageSeverity must be a valid DebugUtilsMessageSeverityFlagBitsEXT value

messageTypes must be a valid combination of DebugUtilsMessageTypeFlagBitsEXT values

messageTypes must not be 0

pCallbackData must be a valid pointer to a valid DebugUtilsMessengerCallbackDataEXT structure

See Also

DebugUtilsMessageSeverityFlagBitsEXT, DebugUtilsMessageTypeFlagsEXT, DebugUtilsMessengerCallbackDataEXT, Instance

data DebugUtilsObjectNameInfoEXT

VkDebugUtilsObjectNameInfoEXT - Specify parameters of a name to give to an object

Description

Applications may change the name associated with an object simply by calling setDebugUtilsObjectNameEXT again with a new string. If pObjectName is either NULL or an empty string, then any previously set name is removed.

Valid Usage

If objectType is OBJECT_TYPE_UNKNOWN, objectHandle must not be NULL_HANDLE

If objectType is not OBJECT_TYPE_UNKNOWN, objectHandle must be NULL_HANDLE or a valid Vulkan handle of the type associated with objectType as defined in the and Vulkan Handle Relationship table

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_DEBUG_UTILS_OBJECT_NAME_INFO_EXT

pNext must be NULL

objectType must be a valid ObjectType value

If pObjectName is not NULL, pObjectName must be a null-terminated UTF-8 string

See Also

DebugUtilsMessengerCallbackDataEXT, ObjectType, StructureType, setDebugUtilsObjectNameEXT

Constructors

DebugUtilsObjectNameInfoEXT

Fields

objectType :: ObjectType

objectType is a ObjectType specifying the type of the object to be named.

objectHandle :: Word64

objectHandle is the object to be named.

objectName :: Maybe ByteString

pObjectName is either NULL or a null-terminated UTF-8 string specifying the name to apply to objectHandle.

data DebugUtilsObjectTagInfoEXT

VkDebugUtilsObjectTagInfoEXT - Specify parameters of a tag to attach to an object

Description

The tagName parameter gives a name or identifier to the type of data being tagged. This can be used by debugging layers to easily filter for only data that can be used by that implementation.

Valid Usage (Implicit)

See Also

ObjectType, StructureType, setDebugUtilsObjectTagEXT

Constructors

DebugUtilsObjectTagInfoEXT

Fields

objectType :: ObjectType

objectType is a ObjectType specifying the type of the object to be named.

objectType must not be OBJECT_TYPE_UNKNOWN

objectType must be a valid ObjectType value

objectHandle :: Word64

objectHandle is the object to be tagged.

objectHandle must be a valid Vulkan handle of the type associated with objectType as defined in the and Vulkan Handle Relationship table

tagName :: Word64

tagName is a numerical identifier of the tag.

tagSize :: Word64

tagSize is the number of bytes of data to attach to the object.

tagSize must be greater than 0

tag :: Ptr ()

pTag is a pointer to an array of tagSize bytes containing the data to be associated with the object.

pTag must be a valid pointer to an array of tagSize bytes

data DebugUtilsLabelEXT

VkDebugUtilsLabelEXT - Specify parameters of a label region

Valid Usage (Implicit)

See Also

DebugUtilsMessengerCallbackDataEXT, StructureType, cmdBeginDebugUtilsLabelEXT, cmdInsertDebugUtilsLabelEXT, queueBeginDebugUtilsLabelEXT, queueInsertDebugUtilsLabelEXT

Constructors

DebugUtilsLabelEXT

Fields

labelName :: ByteString

pLabelName is a pointer to a null-terminated UTF-8 string containing the name of the label.

pLabelName must be a null-terminated UTF-8 string

color :: (Float, Float, Float, Float)

color is an optional RGBA color value that can be associated with the label. A particular implementation may choose to ignore this color value. The values contain RGBA values in order, in the range 0.0 to 1.0. If all elements in color are set to 0.0 then it is ignored.

data DebugUtilsMessengerCreateInfoEXT

VkDebugUtilsMessengerCreateInfoEXT - Structure specifying parameters of a newly created debug messenger

Description

For each DebugUtilsMessengerEXT that is created the DebugUtilsMessengerCreateInfoEXT::messageSeverity and DebugUtilsMessengerCreateInfoEXT::messageType determine when that DebugUtilsMessengerCreateInfoEXT::pfnUserCallback is called. The process to determine if the user’s pfnUserCallback is triggered when an event occurs is as follows:

The implementation will perform a bitwise AND of the event’s DebugUtilsMessageSeverityFlagBitsEXT with the messageSeverity provided during creation of the DebugUtilsMessengerEXT object.

If the value is 0, the message is skipped.

The implementation will perform bitwise AND of the event’s DebugUtilsMessageTypeFlagBitsEXT with the messageType provided during the creation of the DebugUtilsMessengerEXT object.

If the value is 0, the message is skipped.

The callback will trigger a debug message for the current event

The callback will come directly from the component that detected the event, unless some other layer intercepts the calls for its own purposes (filter them in a different way, log to a system error log, etc.).

An application can receive multiple callbacks if multiple DebugUtilsMessengerEXT objects are created. A callback will always be executed in the same thread as the originating Vulkan call.

A callback can be called from multiple threads simultaneously (if the application is making Vulkan calls from multiple threads).

Valid Usage (Implicit)

See Also

PFN_vkDebugUtilsMessengerCallbackEXT, DebugUtilsMessageSeverityFlagsEXT, DebugUtilsMessageTypeFlagsEXT, DebugUtilsMessengerCreateFlagsEXT, StructureType, createDebugUtilsMessengerEXT

Constructors

DebugUtilsMessengerCreateInfoEXT

Fields

flags :: DebugUtilsMessengerCreateFlagsEXT

flags is 0 and is reserved for future use.

flags must be 0

messageSeverity :: DebugUtilsMessageSeverityFlagsEXT

messageSeverity is a bitmask of DebugUtilsMessageSeverityFlagBitsEXT specifying which severity of event(s) will cause this callback to be called.

messageSeverity must be a valid combination of DebugUtilsMessageSeverityFlagBitsEXT values

messageSeverity must not be 0

messageType :: DebugUtilsMessageTypeFlagsEXT

messageType is a bitmask of DebugUtilsMessageTypeFlagBitsEXT specifying which type of event(s) will cause this callback to be called.

messageType must be a valid combination of DebugUtilsMessageTypeFlagBitsEXT values

messageType must not be 0

pfnUserCallback :: PFN_vkDebugUtilsMessengerCallbackEXT

pfnUserCallback is the application callback function to call.

pfnUserCallback must be a valid PFN_vkDebugUtilsMessengerCallbackEXT

pfnUserCallback must be a valid PFN_vkDebugUtilsMessengerCallbackEXT value

userData :: Ptr ()

pUserData is user data to be passed to the callback.

data DebugUtilsMessengerCallbackDataEXT

VkDebugUtilsMessengerCallbackDataEXT - Structure specifying parameters returned to the callback

Description

Note

This structure should only be considered valid during the lifetime of the triggered callback.

Since adding queue and command buffer labels behaves like pushing and popping onto a stack, the order of both pQueueLabels and pCmdBufLabels is based on the order the labels were defined. The result is that the first label in either pQueueLabels or pCmdBufLabels will be the first defined (and therefore the oldest) while the last label in each list will be the most recent.

Note

pQueueLabels will only be non-NULL if one of the objects in pObjects can be related directly to a defined Queue which has had one or more labels associated with it.

Likewise, pCmdBufLabels will only be non-NULL if one of the objects in pObjects can be related directly to a defined CommandBuffer which has had one or more labels associated with it. Additionally, while command buffer labels allow for beginning and ending across different command buffers, the debug messaging framework cannot guarantee that labels in pCmdBufLables will contain those defined outside of the associated command buffer. This is partially due to the fact that the association of one command buffer with another may not have been defined at the time the debug message is triggered.

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_DEBUG_UTILS_MESSENGER_CALLBACK_DATA_EXT

pNext must be NULL

flags must be 0

If pMessageIdName is not NULL, pMessageIdName must be a null-terminated UTF-8 string

pMessage must be a null-terminated UTF-8 string

If queueLabelCount is not 0, pQueueLabels must be a valid pointer to an array of queueLabelCount valid DebugUtilsLabelEXT structures

If cmdBufLabelCount is not 0, pCmdBufLabels must be a valid pointer to an array of cmdBufLabelCount valid DebugUtilsLabelEXT structures

If objectCount is not 0, pObjects must be a valid pointer to an array of objectCount valid DebugUtilsObjectNameInfoEXT structures

See Also

DebugUtilsLabelEXT, DebugUtilsMessengerCallbackDataFlagsEXT, DebugUtilsObjectNameInfoEXT, StructureType, submitDebugUtilsMessageEXT

Constructors

DebugUtilsMessengerCallbackDataEXT

Fields

flags :: DebugUtilsMessengerCallbackDataFlagsEXT

flags is 0 and is reserved for future use.

messageIdName :: Maybe ByteString

pMessageIdName is a null-terminated string that identifies the particular message ID that is associated with the provided message. If the message corresponds to a validation layer message, then this string may contain the portion of the Vulkan specification that is believed to have been violated.

messageIdNumber :: Int32

messageIdNumber is the ID number of the triggering message. If the message corresponds to a validation layer message, then this number is related to the internal number associated with the message being triggered.

message :: ByteString

pMessage is a null-terminated string detailing the trigger conditions.

queueLabels :: Vector DebugUtilsLabelEXT

pQueueLabels is NULL or a pointer to an array of DebugUtilsLabelEXT active in the current Queue at the time the callback was triggered. Refer to Queue Labels for more information.

cmdBufLabels :: Vector DebugUtilsLabelEXT

pCmdBufLabels is NULL or a pointer to an array of DebugUtilsLabelEXT active in the current CommandBuffer at the time the callback was triggered. Refer to Command Buffer Labels for more information.

objects :: Vector DebugUtilsObjectNameInfoEXT

pObjects is a pointer to an array of DebugUtilsObjectNameInfoEXT objects related to the detected issue. The array is roughly in order or importance, but the 0th element is always guaranteed to be the most important object for this message.

newtype DebugUtilsMessengerCreateFlagsEXT

VkDebugUtilsMessengerCreateFlagsEXT - Reserved for future use

Description

DebugUtilsMessengerCreateFlagsEXT is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

DebugUtilsMessengerCreateInfoEXT

Constructors

DebugUtilsMessengerCreateFlagsEXT Flags

newtype DebugUtilsMessengerCallbackDataFlagsEXT

VkDebugUtilsMessengerCallbackDataFlagsEXT - Reserved for future use

Description

DebugUtilsMessengerCallbackDataFlagsEXT is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

DebugUtilsMessengerCallbackDataEXT

Constructors

DebugUtilsMessengerCallbackDataFlagsEXT Flags

newtype DebugUtilsMessageSeverityFlagBitsEXT

VkDebugUtilsMessageSeverityFlagBitsEXT - Bitmask specifying which severities of events cause a debug messenger callback

See Also

DebugUtilsMessageSeverityFlagsEXT, submitDebugUtilsMessageEXT

Constructors

DebugUtilsMessageSeverityFlagBitsEXT Flags

Bundled Patterns

pattern DEBUG_UTILS_MESSAGE_SEVERITY_VERBOSE_BIT_EXT :: DebugUtilsMessageSeverityFlagBitsEXT
DEBUG_UTILS_MESSAGE_SEVERITY_VERBOSE_BIT_EXT specifies the most verbose output indicating all diagnostic messages from the Vulkan loader, layers, and drivers should be captured.

pattern DEBUG_UTILS_MESSAGE_SEVERITY_INFO_BIT_EXT :: DebugUtilsMessageSeverityFlagBitsEXT
DEBUG_UTILS_MESSAGE_SEVERITY_INFO_BIT_EXT specifies an informational message such as resource details that may be handy when debugging an application.

pattern DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT :: DebugUtilsMessageSeverityFlagBitsEXT
DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT specifies use of Vulkan that may expose an app bug. Such cases may not be immediately harmful, such as a fragment shader outputting to a location with no attachment. Other cases may point to behavior that is almost certainly bad when unintended such as using an image whose memory has not been filled. In general if you see a warning but you know that the behavior is intended/desired, then simply ignore the warning.

pattern DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT :: DebugUtilsMessageSeverityFlagBitsEXT
DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT specifies that the application has violated a valid usage condition of the specification.

type DebugUtilsMessageSeverityFlagsEXT = DebugUtilsMessageSeverityFlagBitsEXT

newtype DebugUtilsMessageTypeFlagBitsEXT

VkDebugUtilsMessageTypeFlagBitsEXT - Bitmask specifying which types of events cause a debug messenger callback

See Also

DebugUtilsMessageTypeFlagsEXT

Constructors

DebugUtilsMessageTypeFlagBitsEXT Flags

Bundled Patterns

pattern DEBUG_UTILS_MESSAGE_TYPE_GENERAL_BIT_EXT :: DebugUtilsMessageTypeFlagBitsEXT
DEBUG_UTILS_MESSAGE_TYPE_GENERAL_BIT_EXT specifies that some general event has occurred. This is typically a non-specification, non-performance event.

pattern DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT :: DebugUtilsMessageTypeFlagBitsEXT
DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT specifies that something has occurred during validation against the Vulkan specification that may indicate invalid behavior.

pattern DEBUG_UTILS_MESSAGE_TYPE_PERFORMANCE_BIT_EXT :: DebugUtilsMessageTypeFlagBitsEXT
DEBUG_UTILS_MESSAGE_TYPE_PERFORMANCE_BIT_EXT specifies a potentially non-optimal use of Vulkan, e.g. using cmdClearColorImage when setting AttachmentDescription::loadOp to ATTACHMENT_LOAD_OP_CLEAR would have worked.

type DebugUtilsMessageTypeFlagsEXT = DebugUtilsMessageTypeFlagBitsEXT

type PFN_vkDebugUtilsMessengerCallbackEXT = FunPtr FN_vkDebugUtilsMessengerCallbackEXT

PFN_vkDebugUtilsMessengerCallbackEXT - Application-defined debug messenger callback function

Description

The callback must not call destroyDebugUtilsMessengerEXT.

The callback returns a Bool32, which is interpreted in a layer-specified manner. The application should always return FALSE. The TRUE value is reserved for use in layer development.

See Also

DebugUtilsMessengerCreateInfoEXT

type FN_vkDebugUtilsMessengerCallbackEXT = DebugUtilsMessageSeverityFlagBitsEXT -> ("messageTypes" ::: DebugUtilsMessageTypeFlagsEXT) -> ("pCallbackData" ::: Ptr DebugUtilsMessengerCallbackDataEXT) -> ("pUserData" ::: Ptr ()) -> IO Bool32

type EXT_DEBUG_UTILS_SPEC_VERSION = 2

pattern EXT_DEBUG_UTILS_SPEC_VERSION :: forall a. Integral a => a

type EXT_DEBUG_UTILS_EXTENSION_NAME = "VK_EXT_debug_utils"

pattern EXT_DEBUG_UTILS_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype DebugUtilsMessengerEXT

VkDebugUtilsMessengerEXT - Opaque handle to a debug messenger object

Description

The debug messenger will provide detailed feedback on the application’s use of Vulkan when events of interest occur. When an event of interest does occur, the debug messenger will submit a debug message to the debug callback that was provided during its creation. Additionally, the debug messenger is responsible with filtering out debug messages that the callback is not interested in and will only provide desired debug messages.

See Also

createDebugUtilsMessengerEXT, destroyDebugUtilsMessengerEXT

Constructors

DebugUtilsMessengerEXT Word64