Vulkan.Extensions.VK_EXT_depth_clip_enable

Documentation

data PhysicalDeviceDepthClipEnableFeaturesEXT

VkPhysicalDeviceDepthClipEnableFeaturesEXT - Structure indicating support for explicit enable of depth clip

Members

The members of the PhysicalDeviceDepthClipEnableFeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceDepthClipEnableFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether the feature is supported. PhysicalDeviceDepthClipEnableFeaturesEXT can also be included in the pNext chain of DeviceCreateInfo to enable this feature.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceDepthClipEnableFeaturesEXT

Fields

depthClipEnable :: Bool

depthClipEnable indicates that the implementation supports setting the depth clipping operation explicitly via the PipelineRasterizationDepthClipStateCreateInfoEXT pipeline state. Otherwise depth clipping is only enabled when PipelineRasterizationStateCreateInfo::depthClampEnable is set to FALSE.

data PipelineRasterizationDepthClipStateCreateInfoEXT

VkPipelineRasterizationDepthClipStateCreateInfoEXT - Structure specifying depth clipping state

Valid Usage (Implicit)

See Also

Bool32, PipelineRasterizationDepthClipStateCreateFlagsEXT, StructureType

Constructors

PipelineRasterizationDepthClipStateCreateInfoEXT

Fields

flags :: PipelineRasterizationDepthClipStateCreateFlagsEXT

flags is reserved for future use.

flags must be 0

depthClipEnable :: Bool

depthClipEnable controls whether depth clipping is enabled as described in Primitive Clipping.

newtype PipelineRasterizationDepthClipStateCreateFlagsEXT

VkPipelineRasterizationDepthClipStateCreateFlagsEXT - Reserved for future use

Description

PipelineRasterizationDepthClipStateCreateFlagsEXT is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

PipelineRasterizationDepthClipStateCreateInfoEXT

Constructors

PipelineRasterizationDepthClipStateCreateFlagsEXT Flags

type EXT_DEPTH_CLIP_ENABLE_SPEC_VERSION = 1

pattern EXT_DEPTH_CLIP_ENABLE_SPEC_VERSION :: forall a. Integral a => a

type EXT_DEPTH_CLIP_ENABLE_EXTENSION_NAME = "VK_EXT_depth_clip_enable"

pattern EXT_DEPTH_CLIP_ENABLE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a