Vulkan.Extensions.VK_EXT_device_memory_report

Documentation

data PhysicalDeviceDeviceMemoryReportFeaturesEXT

VkPhysicalDeviceDeviceMemoryReportFeaturesEXT - Structure describing whether device memory report callback can be supported by an implementation

Members

The members of the PhysicalDeviceDeviceMemoryReportFeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceDeviceMemoryReportFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with a value indicating whether the feature is supported. PhysicalDeviceDeviceMemoryReportFeaturesEXT can also be used in the pNext chain of DeviceCreateInfo to enable the feature.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceDeviceMemoryReportFeaturesEXT

Fields

deviceMemoryReport :: Bool

deviceMemoryReport indicates whether the implementation supports the ability to register device memory report callbacks.

data DeviceDeviceMemoryReportCreateInfoEXT

VkDeviceDeviceMemoryReportCreateInfoEXT - Register device memory report callbacks for a Vulkan device

Description

The callback may be called from multiple threads simultaneously.

The callback must be called only once by the implementation when a DeviceMemoryReportEventTypeEXT event occurs.

Note

The callback could be called from a background thread other than the thread calling the Vulkan commands.

Valid Usage (Implicit)

See Also

PFN_vkDeviceMemoryReportCallbackEXT, DeviceMemoryReportFlagsEXT, StructureType

Constructors

DeviceDeviceMemoryReportCreateInfoEXT

Fields

flags :: DeviceMemoryReportFlagsEXT

flags is 0 and reserved for future use.

flags must be 0

pfnUserCallback :: PFN_vkDeviceMemoryReportCallbackEXT

pfnUserCallback is the application callback function to call.

pfnUserCallback must be a valid PFN_vkDeviceMemoryReportCallbackEXT value

userData :: Ptr ()

pUserData is user data to be passed to the callback.

pUserData must be a pointer value

data DeviceMemoryReportCallbackDataEXT

VkDeviceMemoryReportCallbackDataEXT - Structure specifying parameters returned to the callback

Description

memoryObjectId is used to avoid double-counting on the same memory object.

If an internally-allocated device memory object or a DeviceMemory cannot be exported, memoryObjectId must be unique in the Device.

If an internally-allocated device memory object or a DeviceMemory supports being exported, memoryObjectId must be unique system wide.

If an internal device memory object or a DeviceMemory is backed by an imported external memory object, memoryObjectId must be unique system wide.

Note

This structure should only be considered valid during the lifetime of the triggered callback.

For DEVICE_MEMORY_REPORT_EVENT_TYPE_ALLOCATE_EXT and DEVICE_MEMORY_REPORT_EVENT_TYPE_IMPORT_EXT events, objectHandle usually will not yet exist when the application or tool receives the callback. objectHandle will only exist when the create or allocate call that triggered the event returns, and if the allocation or import ends up failing objectHandle won’t ever exist.

Valid Usage (Implicit)

See Also

DeviceMemoryReportEventTypeEXT, DeviceMemoryReportFlagsEXT, DeviceSize, ObjectType, StructureType

Constructors

DeviceMemoryReportCallbackDataEXT

Fields

flags :: DeviceMemoryReportFlagsEXT

flags is 0 and reserved for future use.

type' :: DeviceMemoryReportEventTypeEXT

type is a DeviceMemoryReportEventTypeEXT type specifying the type of event reported in this DeviceMemoryReportCallbackDataEXT structure.

memoryObjectId :: Word64

memoryObjectId is the unique id for the underlying memory object as described below.

size :: DeviceSize

size is the size of the memory object in bytes. If type is DEVICE_MEMORY_REPORT_EVENT_TYPE_ALLOCATE_EXT, DEVICE_MEMORY_REPORT_EVENT_TYPE_IMPORT_EXT or DEVICE_MEMORY_REPORT_EVENT_TYPE_ALLOCATION_FAILED_EXT, size must be a valid DeviceSize value.

objectType :: ObjectType

objectType is a ObjectType value specifying the type of the object associated with this device memory report event. If type is DEVICE_MEMORY_REPORT_EVENT_TYPE_ALLOCATE_EXT, DEVICE_MEMORY_REPORT_EVENT_TYPE_IMPORT_EXT or DEVICE_MEMORY_REPORT_EVENT_TYPE_ALLOCATION_FAILED_EXT, objectType must be a valid ObjectType enum.

objectHandle :: Word64

objectHandle is the object this device memory report event is attributed to. If type is DEVICE_MEMORY_REPORT_EVENT_TYPE_ALLOCATE_EXT, DEVICE_MEMORY_REPORT_EVENT_TYPE_FREE_EXT, DEVICE_MEMORY_REPORT_EVENT_TYPE_IMPORT_EXT or DEVICE_MEMORY_REPORT_EVENT_TYPE_UNIMPORT_EXT, objectHandle must be a valid Vulkan handle of the type associated with objectType as defined in the VkObjectType and Vulkan Handle Relationship table.

heapIndex :: Word32

heapIndex describes which memory heap this device memory allocation is made from. If type is DEVICE_MEMORY_REPORT_EVENT_TYPE_ALLOCATE_EXT or DEVICE_MEMORY_REPORT_EVENT_TYPE_ALLOCATION_FAILED_EXT, heapIndex must correspond to one of the valid heaps from the PhysicalDeviceMemoryProperties structure. Otherwise, heapIndex is undefined.

newtype DeviceMemoryReportFlagsEXT

Constructors

DeviceMemoryReportFlagsEXT Flags

newtype DeviceMemoryReportEventTypeEXT

VkDeviceMemoryReportEventTypeEXT - Events that can occur on a device memory object

See Also

DeviceMemoryReportCallbackDataEXT

Constructors

DeviceMemoryReportEventTypeEXT Int32

Bundled Patterns

pattern DEVICE_MEMORY_REPORT_EVENT_TYPE_ALLOCATE_EXT :: DeviceMemoryReportEventTypeEXT
DEVICE_MEMORY_REPORT_EVENT_TYPE_ALLOCATE_EXT specifies this event corresponds to the allocation of an internal device memory object or a DeviceMemory.

pattern DEVICE_MEMORY_REPORT_EVENT_TYPE_FREE_EXT :: DeviceMemoryReportEventTypeEXT
DEVICE_MEMORY_REPORT_EVENT_TYPE_FREE_EXT specifies this event corresponds to the deallocation of an internally-allocated device memory object or a DeviceMemory.

pattern DEVICE_MEMORY_REPORT_EVENT_TYPE_IMPORT_EXT :: DeviceMemoryReportEventTypeEXT
DEVICE_MEMORY_REPORT_EVENT_TYPE_IMPORT_EXT specifies this event corresponds to the import of an external memory object.

pattern DEVICE_MEMORY_REPORT_EVENT_TYPE_UNIMPORT_EXT :: DeviceMemoryReportEventTypeEXT
DEVICE_MEMORY_REPORT_EVENT_TYPE_UNIMPORT_EXT specifies this event is the release of an imported external memory object.

pattern DEVICE_MEMORY_REPORT_EVENT_TYPE_ALLOCATION_FAILED_EXT :: DeviceMemoryReportEventTypeEXT
DEVICE_MEMORY_REPORT_EVENT_TYPE_ALLOCATION_FAILED_EXT specifies this event corresponds to the failed allocation of an internal device memory object or a DeviceMemory.

type PFN_vkDeviceMemoryReportCallbackEXT = FunPtr FN_vkDeviceMemoryReportCallbackEXT

PFN_vkDeviceMemoryReportCallbackEXT - Application-defined device memory report callback function

Description

The callback must not make calls to any Vulkan commands.

See Also

DeviceDeviceMemoryReportCreateInfoEXT

type FN_vkDeviceMemoryReportCallbackEXT = ("pCallbackData" ::: Ptr DeviceMemoryReportCallbackDataEXT) -> ("pUserData" ::: Ptr ()) -> IO ()

type EXT_DEVICE_MEMORY_REPORT_SPEC_VERSION = 1

pattern EXT_DEVICE_MEMORY_REPORT_SPEC_VERSION :: forall a. Integral a => a

type EXT_DEVICE_MEMORY_REPORT_EXTENSION_NAME = "VK_EXT_device_memory_report"

pattern EXT_DEVICE_MEMORY_REPORT_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a