Vulkan.Extensions.VK_EXT_external_memory_host

Documentation

getMemoryHostPointerPropertiesEXT

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that will be importing pHostPointer.

-> ExternalMemoryHandleTypeFlagBits
handleType is the type of the handle pHostPointer.

-> ("hostPointer" ::: Ptr ())
pHostPointer is the host pointer to import from.

-> io MemoryHostPointerPropertiesEXT

vkGetMemoryHostPointerPropertiesEXT - Get properties of external memory host pointer

Valid Usage

handleType must be EXTERNAL_MEMORY_HANDLE_TYPE_HOST_ALLOCATION_BIT_EXT or EXTERNAL_MEMORY_HANDLE_TYPE_HOST_MAPPED_FOREIGN_MEMORY_BIT_EXT

pHostPointer must be a pointer aligned to an integer multiple of PhysicalDeviceExternalMemoryHostPropertiesEXT::minImportedHostPointerAlignment

If handleType is EXTERNAL_MEMORY_HANDLE_TYPE_HOST_ALLOCATION_BIT_EXT, pHostPointer must be a pointer to host memory

If handleType is EXTERNAL_MEMORY_HANDLE_TYPE_HOST_MAPPED_FOREIGN_MEMORY_BIT_EXT, pHostPointer must be a pointer to host mapped foreign memory

Valid Usage (Implicit)

device must be a valid Device handle

handleType must be a valid ExternalMemoryHandleTypeFlagBits value

pMemoryHostPointerProperties must be a valid pointer to a MemoryHostPointerPropertiesEXT structure

Return Codes

Success

SUCCESS

Failure

ERROR_INVALID_EXTERNAL_HANDLE

See Also

Device, ExternalMemoryHandleTypeFlagBits, MemoryHostPointerPropertiesEXT

data ImportMemoryHostPointerInfoEXT

VkImportMemoryHostPointerInfoEXT - import memory from a host pointer

Description

Importing memory from a host pointer shares ownership of the memory between the host and the Vulkan implementation. The application can continue to access the memory through the host pointer but it is the application’s responsibility to synchronize device and non-device access to the underlying memory as defined in Host Access to Device Memory Objects.

Applications can import the same underlying memory into multiple instances of Vulkan and multiple times into a given Vulkan instance. However, implementations may fail to import the same underlying memory multiple times into a given physical device due to platform constraints.

Importing memory from a particular host pointer may not be possible due to additional platform-specific restrictions beyond the scope of this specification in which case the implementation must fail the memory import operation with the error code ERROR_INVALID_EXTERNAL_HANDLE_KHR.

The application must ensure that the imported memory range remains valid and accessible for the lifetime of the imported memory object.

Valid Usage

If handleType is not 0, it must be supported for import, as reported in ExternalMemoryProperties

If handleType is not 0, it must be EXTERNAL_MEMORY_HANDLE_TYPE_HOST_ALLOCATION_BIT_EXT or EXTERNAL_MEMORY_HANDLE_TYPE_HOST_MAPPED_FOREIGN_MEMORY_BIT_EXT

pHostPointer must be a pointer aligned to an integer multiple of PhysicalDeviceExternalMemoryHostPropertiesEXT::minImportedHostPointerAlignment

If handleType is EXTERNAL_MEMORY_HANDLE_TYPE_HOST_ALLOCATION_BIT_EXT, pHostPointer must be a pointer to allocationSize number of bytes of host memory, where allocationSize is the member of the MemoryAllocateInfo structure this structure is chained to

If handleType is EXTERNAL_MEMORY_HANDLE_TYPE_HOST_MAPPED_FOREIGN_MEMORY_BIT_EXT, pHostPointer must be a pointer to allocationSize number of bytes of host mapped foreign memory, where allocationSize is the member of the MemoryAllocateInfo structure this structure is chained to

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_IMPORT_MEMORY_HOST_POINTER_INFO_EXT

handleType must be a valid ExternalMemoryHandleTypeFlagBits value

See Also

ExternalMemoryHandleTypeFlagBits, StructureType

Constructors

ImportMemoryHostPointerInfoEXT

Fields

handleType :: ExternalMemoryHandleTypeFlagBits

handleType specifies the handle type.

hostPointer :: Ptr ()

pHostPointer is the host pointer to import from.

data MemoryHostPointerPropertiesEXT

VkMemoryHostPointerPropertiesEXT - Properties of external memory host pointer

Description

The value returned by memoryTypeBits must only include bits that identify memory types which are host visible.

Valid Usage (Implicit)

See Also

StructureType, getMemoryHostPointerPropertiesEXT

Constructors

MemoryHostPointerPropertiesEXT

Fields

memoryTypeBits :: Word32

memoryTypeBits is a bitmask containing one bit set for every memory type which the specified host pointer can be imported as.

data PhysicalDeviceExternalMemoryHostPropertiesEXT

VkPhysicalDeviceExternalMemoryHostPropertiesEXT - Structure describing external memory host pointer limits that can be supported by an implementation

Members

The members of the PhysicalDeviceExternalMemoryHostPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the PhysicalDeviceExternalMemoryHostPropertiesEXT structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

DeviceSize, StructureType

Constructors

PhysicalDeviceExternalMemoryHostPropertiesEXT

Fields

minImportedHostPointerAlignment :: DeviceSize

minImportedHostPointerAlignment is the minimum required alignment, in bytes, for the base address and size of host pointers that can be imported to a Vulkan memory object.

type EXT_EXTERNAL_MEMORY_HOST_SPEC_VERSION = 1

pattern EXT_EXTERNAL_MEMORY_HOST_SPEC_VERSION :: forall a. Integral a => a

type EXT_EXTERNAL_MEMORY_HOST_EXTENSION_NAME = "VK_EXT_external_memory_host"

pattern EXT_EXTERNAL_MEMORY_HOST_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a