Vulkan.Extensions.VK_EXT_inline_uniform_block

Documentation

data PhysicalDeviceInlineUniformBlockFeaturesEXT

VkPhysicalDeviceInlineUniformBlockFeaturesEXT - Structure describing inline uniform block features that can be supported by an implementation

Members

The members of the PhysicalDeviceInlineUniformBlockFeaturesEXT structure describe the following features:

Description

If the PhysicalDeviceInlineUniformBlockFeaturesEXT structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether each feature is supported. PhysicalDeviceInlineUniformBlockFeaturesEXT can also be included in the pNext chain of DeviceCreateInfo to enable features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDeviceInlineUniformBlockFeaturesEXT

Fields

inlineUniformBlock :: Bool

inlineUniformBlock indicates whether the implementation supports inline uniform block descriptors. If this feature is not enabled, DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT must not be used.

descriptorBindingInlineUniformBlockUpdateAfterBind :: Bool

descriptorBindingInlineUniformBlockUpdateAfterBind indicates whether the implementation supports updating inline uniform block descriptors after a set is bound. If this feature is not enabled, DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT must not be used with DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT.

data PhysicalDeviceInlineUniformBlockPropertiesEXT

VkPhysicalDeviceInlineUniformBlockPropertiesEXT - Structure describing inline uniform block properties that can be supported by an implementation

Members

The members of the PhysicalDeviceInlineUniformBlockPropertiesEXT structure describe the following implementation-dependent limits:

Description

If the PhysicalDeviceInlineUniformBlockPropertiesEXT structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

StructureType

Constructors

PhysicalDeviceInlineUniformBlockPropertiesEXT

Fields

maxInlineUniformBlockSize :: Word32

maxInlineUniformBlockSize is the maximum size in bytes of an inline uniform block binding.

maxPerStageDescriptorInlineUniformBlocks :: Word32

maxPerStageDescriptorUpdateAfterBindInlineUniformBlocks :: Word32

maxPerStageDescriptorUpdateAfterBindInlineUniformBlocks is similar to maxPerStageDescriptorInlineUniformBlocks but counts descriptor bindings from descriptor sets created with or without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set.

maxDescriptorSetInlineUniformBlocks :: Word32

maxDescriptorSetInlineUniformBlocks is the maximum number of inline uniform block bindings that can be included in descriptor bindings in a pipeline layout across all pipeline shader stages and descriptor set numbers. Descriptor bindings with a descriptor type of DESCRIPTOR_TYPE_INLINE_UNIFORM_BLOCK_EXT count against this limit. Only descriptor bindings in descriptor set layouts created without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set count against this limit.

maxDescriptorSetUpdateAfterBindInlineUniformBlocks :: Word32

maxDescriptorSetUpdateAfterBindInlineUniformBlocks is similar to maxDescriptorSetInlineUniformBlocks but counts descriptor bindings from descriptor sets created with or without the DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT bit set.

data WriteDescriptorSetInlineUniformBlockEXT

VkWriteDescriptorSetInlineUniformBlockEXT - Structure specifying inline uniform block data

Valid Usage (Implicit)

See Also

StructureType

Constructors

WriteDescriptorSetInlineUniformBlockEXT

Fields

dataSize :: Word32

dataSize is the number of bytes of inline uniform block data pointed to by pData.

dataSize must be an integer multiple of 4

dataSize must be greater than 0

data' :: Ptr ()

pData is a pointer to dataSize number of bytes of data to write to the inline uniform block.

pData must be a valid pointer to an array of dataSize bytes

data DescriptorPoolInlineUniformBlockCreateInfoEXT

VkDescriptorPoolInlineUniformBlockCreateInfoEXT - Structure specifying the maximum number of inline uniform block bindings of a newly created descriptor pool

Valid Usage (Implicit)

See Also

StructureType

Constructors

DescriptorPoolInlineUniformBlockCreateInfoEXT

Fields

maxInlineUniformBlockBindings :: Word32

maxInlineUniformBlockBindings is the number of inline uniform block bindings to allocate.

type EXT_INLINE_UNIFORM_BLOCK_SPEC_VERSION = 1

pattern EXT_INLINE_UNIFORM_BLOCK_SPEC_VERSION :: forall a. Integral a => a

type EXT_INLINE_UNIFORM_BLOCK_EXTENSION_NAME = "VK_EXT_inline_uniform_block"

pattern EXT_INLINE_UNIFORM_BLOCK_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a