Vulkan.Extensions.VK_FUCHSIA_imagepipe_surface

Documentation

createImagePipeSurfaceFUCHSIA

Arguments

:: forall io. MonadIO io

=> Instance
instance is the instance to associate with the surface.

-> ImagePipeSurfaceCreateInfoFUCHSIA
pCreateInfo is a pointer to a ImagePipeSurfaceCreateInfoFUCHSIA structure containing parameters affecting the creation of the surface object.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

-> io SurfaceKHR

vkCreateImagePipeSurfaceFUCHSIA - Create a SurfaceKHR object for a Fuchsia ImagePipe

Valid Usage (Implicit)

instance must be a valid Instance handle

pCreateInfo must be a valid pointer to a valid ImagePipeSurfaceCreateInfoFUCHSIA structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pSurface must be a valid pointer to a SurfaceKHR handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, ImagePipeSurfaceCreateInfoFUCHSIA, Instance, SurfaceKHR

data ImagePipeSurfaceCreateInfoFUCHSIA

VkImagePipeSurfaceCreateInfoFUCHSIA - Structure specifying parameters of a newly created ImagePipe surface object

Valid Usage (Implicit)

See Also

ImagePipeSurfaceCreateFlagsFUCHSIA, StructureType, createImagePipeSurfaceFUCHSIA

Constructors

ImagePipeSurfaceCreateInfoFUCHSIA

Fields

flags :: ImagePipeSurfaceCreateFlagsFUCHSIA

flags is reserved for future use.

flags must be 0

imagePipeHandle :: Zx_handle_t

imagePipeHandle is a zx_handle_t referring to the ImagePipe to associate with the surface.

imagePipeHandle must be a valid zx_handle_t

newtype ImagePipeSurfaceCreateFlagsFUCHSIA

VkImagePipeSurfaceCreateFlagsFUCHSIA - Reserved for future use

Description

ImagePipeSurfaceCreateFlagsFUCHSIA is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

ImagePipeSurfaceCreateInfoFUCHSIA

Constructors

ImagePipeSurfaceCreateFlagsFUCHSIA Flags

type FUCHSIA_IMAGEPIPE_SURFACE_SPEC_VERSION = 1

pattern FUCHSIA_IMAGEPIPE_SURFACE_SPEC_VERSION :: forall a. Integral a => a

type FUCHSIA_IMAGEPIPE_SURFACE_EXTENSION_NAME = "VK_FUCHSIA_imagepipe_surface"

pattern FUCHSIA_IMAGEPIPE_SURFACE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

type Zx_handle_t = Word32

newtype SurfaceKHR

VkSurfaceKHR - Opaque handle to a surface object

Description

The VK_KHR_surface extension declares the SurfaceKHR object, and provides a function for destroying SurfaceKHR objects. Separate platform-specific extensions each provide a function for creating a SurfaceKHR object for the respective platform. From the application’s perspective this is an opaque handle, just like the handles of other Vulkan objects.

See Also

PhysicalDeviceSurfaceInfo2KHR, SwapchainCreateInfoKHR, createAndroidSurfaceKHR, createDirectFBSurfaceEXT, createDisplayPlaneSurfaceKHR, createHeadlessSurfaceEXT, createIOSSurfaceMVK, createImagePipeSurfaceFUCHSIA, createMacOSSurfaceMVK, createMetalSurfaceEXT, createStreamDescriptorSurfaceGGP, createViSurfaceNN, createWaylandSurfaceKHR, createWin32SurfaceKHR, createXcbSurfaceKHR, createXlibSurfaceKHR, destroySurfaceKHR, getDeviceGroupSurfacePresentModesKHR, getPhysicalDevicePresentRectanglesKHR, getPhysicalDeviceSurfaceCapabilities2EXT, getPhysicalDeviceSurfaceCapabilitiesKHR, getPhysicalDeviceSurfaceFormatsKHR, getPhysicalDeviceSurfacePresentModesKHR, getPhysicalDeviceSurfaceSupportKHR

Constructors

SurfaceKHR Word64