Vulkan.Extensions.VK_KHR_display

Documentation

getPhysicalDeviceDisplayPropertiesKHR

Arguments

:: forall io. MonadIO io

=> PhysicalDevice
physicalDevice is a physical device.

-> io (Result, "properties" ::: Vector DisplayPropertiesKHR)

vkGetPhysicalDeviceDisplayPropertiesKHR - Query information about the available displays

Description

If pProperties is NULL, then the number of display devices available for physicalDevice is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If the value of pPropertyCount is less than the number of display devices for physicalDevice, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of display devices available for physicalDevice, INCOMPLETE will be returned instead of SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid PhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount DisplayPropertiesKHR structures

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

DisplayPropertiesKHR, PhysicalDevice

getPhysicalDeviceDisplayPlanePropertiesKHR

Arguments

:: forall io. MonadIO io

=> PhysicalDevice
physicalDevice is a physical device.

-> io (Result, "properties" ::: Vector DisplayPlanePropertiesKHR)

vkGetPhysicalDeviceDisplayPlanePropertiesKHR - Query the plane properties

Description

If pProperties is NULL, then the number of display planes available for physicalDevice is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If the value of pPropertyCount is less than the number of display planes for physicalDevice, at most pPropertyCount structures will be written.

Valid Usage (Implicit)

physicalDevice must be a valid PhysicalDevice handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount DisplayPlanePropertiesKHR structures

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

DisplayPlanePropertiesKHR, PhysicalDevice

getDisplayPlaneSupportedDisplaysKHR

Arguments

:: forall io. MonadIO io

=> PhysicalDevice
physicalDevice is a physical device.

-> ("planeIndex" ::: Word32)
planeIndex is the plane which the application wishes to use, and must be in the range [0, physical device plane count - 1].

-> io (Result, "displays" ::: Vector DisplayKHR)

vkGetDisplayPlaneSupportedDisplaysKHR - Query the list of displays a plane supports

Description

If pDisplays is NULL, then the number of displays usable with the specified planeIndex for physicalDevice is returned in pDisplayCount. Otherwise, pDisplayCount must point to a variable set by the user to the number of elements in the pDisplays array, and on return the variable is overwritten with the number of handles actually written to pDisplays. If the value of pDisplayCount is less than the number of display planes for physicalDevice, at most pDisplayCount handles will be written. If pDisplayCount is smaller than the number of displays usable with the specified planeIndex for physicalDevice, INCOMPLETE will be returned instead of SUCCESS to indicate that not all the available values were returned.

Valid Usage

planeIndex must be less than the number of display planes supported by the device as determined by calling getPhysicalDeviceDisplayPlanePropertiesKHR

Valid Usage (Implicit)

physicalDevice must be a valid PhysicalDevice handle

pDisplayCount must be a valid pointer to a uint32_t value

If the value referenced by pDisplayCount is not 0, and pDisplays is not NULL, pDisplays must be a valid pointer to an array of pDisplayCount DisplayKHR handles

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

DisplayKHR, PhysicalDevice

getDisplayModePropertiesKHR

Arguments

:: forall io. MonadIO io

=> PhysicalDevice
physicalDevice is the physical device associated with display.

-> DisplayKHR
display is the display to query.

-> io (Result, "properties" ::: Vector DisplayModePropertiesKHR)

vkGetDisplayModePropertiesKHR - Query the set of mode properties supported by the display

Description

If pProperties is NULL, then the number of display modes available on the specified display for physicalDevice is returned in pPropertyCount. Otherwise, pPropertyCount must point to a variable set by the user to the number of elements in the pProperties array, and on return the variable is overwritten with the number of structures actually written to pProperties. If the value of pPropertyCount is less than the number of display modes for physicalDevice, at most pPropertyCount structures will be written. If pPropertyCount is smaller than the number of display modes available on the specified display for physicalDevice, INCOMPLETE will be returned instead of SUCCESS to indicate that not all the available values were returned.

Valid Usage (Implicit)

physicalDevice must be a valid PhysicalDevice handle

display must be a valid DisplayKHR handle

pPropertyCount must be a valid pointer to a uint32_t value

If the value referenced by pPropertyCount is not 0, and pProperties is not NULL, pProperties must be a valid pointer to an array of pPropertyCount DisplayModePropertiesKHR structures

display must have been created, allocated, or retrieved from physicalDevice

Return Codes

Success

SUCCESS

INCOMPLETE

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

DisplayKHR, DisplayModePropertiesKHR, PhysicalDevice

createDisplayModeKHR

Arguments

:: forall io. MonadIO io

=> PhysicalDevice
physicalDevice is the physical device associated with display.

-> DisplayKHR
display is the display to create an additional mode for.

-> DisplayModeCreateInfoKHR
pCreateInfo is a DisplayModeCreateInfoKHR structure describing the new mode to create.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator is the allocator used for host memory allocated for the display mode object when there is no more specific allocator available (see Memory Allocation).

-> io DisplayModeKHR

vkCreateDisplayModeKHR - Create a display mode

Valid Usage (Implicit)

physicalDevice must be a valid PhysicalDevice handle

display must be a valid DisplayKHR handle

pCreateInfo must be a valid pointer to a valid DisplayModeCreateInfoKHR structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pMode must be a valid pointer to a DisplayModeKHR handle

display must have been created, allocated, or retrieved from physicalDevice

Host Synchronization

Host access to display must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_INITIALIZATION_FAILED

See Also

AllocationCallbacks, DisplayKHR, DisplayModeCreateInfoKHR, DisplayModeKHR, PhysicalDevice

getDisplayPlaneCapabilitiesKHR

Arguments

:: forall io. MonadIO io

=> PhysicalDevice
physicalDevice is the physical device associated with display

-> DisplayModeKHR
mode is the display mode the application intends to program when using the specified plane. Note this parameter also implicitly specifies a display.

-> ("planeIndex" ::: Word32)
planeIndex is the plane which the application intends to use with the display, and is less than the number of display planes supported by the device.

-> io DisplayPlaneCapabilitiesKHR

vkGetDisplayPlaneCapabilitiesKHR - Query capabilities of a mode and plane combination

Valid Usage (Implicit)

physicalDevice must be a valid PhysicalDevice handle

mode must be a valid DisplayModeKHR handle

pCapabilities must be a valid pointer to a DisplayPlaneCapabilitiesKHR structure

Host Synchronization

Host access to mode must be externally synchronized

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

DisplayModeKHR, DisplayPlaneCapabilitiesKHR, PhysicalDevice

createDisplayPlaneSurfaceKHR

Arguments

:: forall io. MonadIO io

=> Instance
instance is the instance corresponding to the physical device the targeted display is on.

-> DisplaySurfaceCreateInfoKHR
pCreateInfo is a pointer to a DisplaySurfaceCreateInfoKHR structure specifying which mode, plane, and other parameters to use, as described below.

-> ("allocator" ::: Maybe AllocationCallbacks)
pAllocator is the allocator used for host memory allocated for the surface object when there is no more specific allocator available (see Memory Allocation).

-> io SurfaceKHR

vkCreateDisplayPlaneSurfaceKHR - Create a SurfaceKHR structure representing a display plane and mode

Valid Usage (Implicit)

instance must be a valid Instance handle

pCreateInfo must be a valid pointer to a valid DisplaySurfaceCreateInfoKHR structure

If pAllocator is not NULL, pAllocator must be a valid pointer to a valid AllocationCallbacks structure

pSurface must be a valid pointer to a SurfaceKHR handle

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

See Also

AllocationCallbacks, DisplaySurfaceCreateInfoKHR, Instance, SurfaceKHR

data DisplayPropertiesKHR

VkDisplayPropertiesKHR - Structure describing an available display device

Description

Note

For devices which have no natural value to return here, implementations should return the maximum resolution supported.

Note

Persistent presents may have higher latency, and may use less power when the screen content is updated infrequently, or when only a portion of the screen needs to be updated in most frames.

See Also

Bool32, DisplayKHR, DisplayProperties2KHR, Extent2D, SurfaceTransformFlagsKHR, getPhysicalDeviceDisplayPropertiesKHR

Constructors

DisplayPropertiesKHR

Fields

display :: DisplayKHR

display is a handle that is used to refer to the display described here. This handle will be valid for the lifetime of the Vulkan instance.

displayName :: ByteString

displayName is a pointer to a null-terminated UTF-8 string containing the name of the display. Generally, this will be the name provided by the display’s EDID. It can be NULL if no suitable name is available. If not NULL, the memory it points to must remain accessible as long as display is valid.

physicalDimensions :: Extent2D

physicalDimensions describes the physical width and height of the visible portion of the display, in millimeters.

physicalResolution :: Extent2D

physicalResolution describes the physical, native, or preferred resolution of the display.

supportedTransforms :: SurfaceTransformFlagsKHR

supportedTransforms is a bitmask of SurfaceTransformFlagBitsKHR describing which transforms are supported by this display.

planeReorderPossible :: Bool

planeReorderPossible tells whether the planes on this display can have their z order changed. If this is TRUE, the application can re-arrange the planes on this display in any order relative to each other.

persistentContent :: Bool

persistentContent tells whether the display supports self-refresh/internal buffering. If this is true, the application can submit persistent present operations on swapchains created against this display.

data DisplayPlanePropertiesKHR

VkDisplayPlanePropertiesKHR - Structure describing display plane properties

See Also

DisplayKHR, DisplayPlaneProperties2KHR, getPhysicalDeviceDisplayPlanePropertiesKHR

Constructors

DisplayPlanePropertiesKHR

Fields

currentDisplay :: DisplayKHR

currentDisplay is the handle of the display the plane is currently associated with. If the plane is not currently attached to any displays, this will be NULL_HANDLE.

currentStackIndex :: Word32

currentStackIndex is the current z-order of the plane. This will be between 0 and the value returned by getPhysicalDeviceDisplayPlanePropertiesKHR in pPropertyCount.

data DisplayModeParametersKHR

VkDisplayModeParametersKHR - Structure describing display parameters associated with a display mode

Description

Note

For example, a 60Hz display mode would report a refreshRate of 60,000.

Valid Usage

The width member of visibleRegion must be greater than 0

The height member of visibleRegion must be greater than 0

refreshRate must be greater than 0

See Also

DisplayModeCreateInfoKHR, DisplayModePropertiesKHR, Extent2D

Constructors

DisplayModeParametersKHR

Fields

visibleRegion :: Extent2D

visibleRegion is the 2D extents of the visible region.

refreshRate :: Word32

refreshRate is a uint32_t that is the number of times the display is refreshed each second multiplied by 1000.

data DisplayModePropertiesKHR

VkDisplayModePropertiesKHR - Structure describing display mode properties

See Also

DisplayModeKHR, DisplayModeParametersKHR, DisplayModeProperties2KHR, getDisplayModePropertiesKHR

Constructors

DisplayModePropertiesKHR

Fields

displayMode :: DisplayModeKHR

displayMode is a handle to the display mode described in this structure. This handle will be valid for the lifetime of the Vulkan instance.

parameters :: DisplayModeParametersKHR

parameters is a DisplayModeParametersKHR structure describing the display parameters associated with displayMode.

data DisplayModeCreateInfoKHR

VkDisplayModeCreateInfoKHR - Structure specifying parameters of a newly created display mode object

Valid Usage (Implicit)

See Also

DisplayModeCreateFlagsKHR, DisplayModeParametersKHR, StructureType, createDisplayModeKHR

Constructors

DisplayModeCreateInfoKHR

Fields

flags :: DisplayModeCreateFlagsKHR

flags is reserved for future use, and must be zero.

flags must be 0

parameters :: DisplayModeParametersKHR

parameters is a DisplayModeParametersKHR structure describing the display parameters to use in creating the new mode. If the parameters are not compatible with the specified display, the implementation must return ERROR_INITIALIZATION_FAILED.

parameters must be a valid DisplayModeParametersKHR structure

data DisplayPlaneCapabilitiesKHR

VkDisplayPlaneCapabilitiesKHR - Structure describing capabilities of a mode and plane combination

Description

The minimum and maximum position and extent fields describe the implementation limits, if any, as they apply to the specified display mode and plane. Vendors may support displaying a subset of a swapchain’s presentable images on the specified display plane. This is expressed by returning minSrcPosition, maxSrcPosition, minSrcExtent, and maxSrcExtent values that indicate a range of possible positions and sizes may be used to specify the region within the presentable images that source pixels will be read from when creating a swapchain on the specified display mode and plane.

Vendors may also support mapping the presentable images’ content to a subset or superset of the visible region in the specified display mode. This is expressed by returning minDstPosition, maxDstPosition, minDstExtent and maxDstExtent values that indicate a range of possible positions and sizes may be used to describe the region within the display mode that the source pixels will be mapped to.

Other vendors may support only a 1-1 mapping between pixels in the presentable images and the display mode. This may be indicated by returning (0,0) for minSrcPosition, maxSrcPosition, minDstPosition, and maxDstPosition, and (display mode width, display mode height) for minSrcExtent, maxSrcExtent, minDstExtent, and maxDstExtent.

These values indicate the limits of the implementation’s individual fields. Not all combinations of values within the offset and extent ranges returned in DisplayPlaneCapabilitiesKHR are guaranteed to be supported. Presentation requests specifying unsupported combinations may fail.

See Also

DisplayPlaneAlphaFlagsKHR, DisplayPlaneCapabilities2KHR, Extent2D, Offset2D, getDisplayPlaneCapabilitiesKHR

Constructors

DisplayPlaneCapabilitiesKHR

Fields

supportedAlpha :: DisplayPlaneAlphaFlagsKHR

supportedAlpha is a bitmask of DisplayPlaneAlphaFlagBitsKHR describing the supported alpha blending modes.

minSrcPosition :: Offset2D

minSrcPosition is the minimum source rectangle offset supported by this plane using the specified mode.

maxSrcPosition :: Offset2D

maxSrcPosition is the maximum source rectangle offset supported by this plane using the specified mode. The x and y components of maxSrcPosition must each be greater than or equal to the x and y components of minSrcPosition, respectively.

minSrcExtent :: Extent2D

minSrcExtent is the minimum source rectangle size supported by this plane using the specified mode.

maxSrcExtent :: Extent2D

maxSrcExtent is the maximum source rectangle size supported by this plane using the specified mode.

minDstPosition :: Offset2D

minDstPosition, maxDstPosition, minDstExtent, maxDstExtent all have similar semantics to their corresponding *Src* equivalents, but apply to the output region within the mode rather than the input region within the source image. Unlike the *Src* offsets, minDstPosition and maxDstPosition may contain negative values.

maxDstPosition :: Offset2D

minDstExtent :: Extent2D

maxDstExtent :: Extent2D

data DisplaySurfaceCreateInfoKHR

VkDisplaySurfaceCreateInfoKHR - Structure specifying parameters of a newly created display plane surface object

Description

Note

Creating a display surface must not modify the state of the displays, planes, or other resources it names. For example, it must not apply the specified mode to be set on the associated display. Application of display configuration occurs as a side effect of presenting to a display surface.

Valid Usage

planeIndex must be less than the number of display planes supported by the device as determined by calling getPhysicalDeviceDisplayPlanePropertiesKHR

If the planeReorderPossible member of the DisplayPropertiesKHR structure returned by getPhysicalDeviceDisplayPropertiesKHR for the display corresponding to displayMode is TRUE then planeStackIndex must be less than the number of display planes supported by the device as determined by calling getPhysicalDeviceDisplayPlanePropertiesKHR; otherwise planeStackIndex must equal the currentStackIndex member of DisplayPlanePropertiesKHR returned by getPhysicalDeviceDisplayPlanePropertiesKHR for the display plane corresponding to displayMode

If alphaMode is DISPLAY_PLANE_ALPHA_GLOBAL_BIT_KHR then globalAlpha must be between 0 and 1, inclusive

alphaMode must be 0 or one of the bits present in the supportedAlpha member of DisplayPlaneCapabilitiesKHR returned by getDisplayPlaneCapabilitiesKHR for the display plane corresponding to displayMode

The width and height members of imageExtent must be less than the maxImageDimensions2D member of PhysicalDeviceLimits

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_DISPLAY_SURFACE_CREATE_INFO_KHR

pNext must be NULL

flags must be 0

displayMode must be a valid DisplayModeKHR handle

transform must be a valid SurfaceTransformFlagBitsKHR value

alphaMode must be a valid DisplayPlaneAlphaFlagBitsKHR value

See Also

DisplayModeKHR, DisplayPlaneAlphaFlagBitsKHR, DisplaySurfaceCreateFlagsKHR, Extent2D, StructureType, SurfaceTransformFlagBitsKHR, createDisplayPlaneSurfaceKHR

Constructors

DisplaySurfaceCreateInfoKHR

Fields

flags :: DisplaySurfaceCreateFlagsKHR

flags is reserved for future use, and must be zero.

displayMode :: DisplayModeKHR

displayMode is a DisplayModeKHR handle specifying the mode to use when displaying this surface.

planeIndex :: Word32

planeIndex is the plane on which this surface appears.

planeStackIndex :: Word32

planeStackIndex is the z-order of the plane.

transform :: SurfaceTransformFlagBitsKHR

transform is a SurfaceTransformFlagBitsKHR value specifying the transformation to apply to images as part of the scanout operation.

globalAlpha :: Float

globalAlpha is the global alpha value. This value is ignored if alphaMode is not DISPLAY_PLANE_ALPHA_GLOBAL_BIT_KHR.

alphaMode :: DisplayPlaneAlphaFlagBitsKHR

alphaMode is a DisplayPlaneAlphaFlagBitsKHR value specifying the type of alpha blending to use.

imageExtent :: Extent2D

imageExtent The size of the presentable images to use with the surface.

newtype DisplayModeCreateFlagsKHR

VkDisplayModeCreateFlagsKHR - Reserved for future use

Description

DisplayModeCreateFlagsKHR is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

DisplayModeCreateInfoKHR

Constructors

DisplayModeCreateFlagsKHR Flags

newtype DisplaySurfaceCreateFlagsKHR

VkDisplaySurfaceCreateFlagsKHR - Reserved for future use

Description

DisplaySurfaceCreateFlagsKHR is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

DisplaySurfaceCreateInfoKHR

Constructors

DisplaySurfaceCreateFlagsKHR Flags

newtype DisplayPlaneAlphaFlagBitsKHR

VkDisplayPlaneAlphaFlagBitsKHR - Alpha blending type

See Also

DisplayPlaneAlphaFlagsKHR, DisplaySurfaceCreateInfoKHR

Constructors

DisplayPlaneAlphaFlagBitsKHR Flags

Bundled Patterns

pattern DISPLAY_PLANE_ALPHA_OPAQUE_BIT_KHR :: DisplayPlaneAlphaFlagBitsKHR
DISPLAY_PLANE_ALPHA_OPAQUE_BIT_KHR specifies that the source image will be treated as opaque.

pattern DISPLAY_PLANE_ALPHA_GLOBAL_BIT_KHR :: DisplayPlaneAlphaFlagBitsKHR
DISPLAY_PLANE_ALPHA_GLOBAL_BIT_KHR specifies that a global alpha value must be specified that will be applied to all pixels in the source image.

pattern DISPLAY_PLANE_ALPHA_PER_PIXEL_BIT_KHR :: DisplayPlaneAlphaFlagBitsKHR
DISPLAY_PLANE_ALPHA_PER_PIXEL_BIT_KHR specifies that the alpha value will be determined by the alpha channel of the source image’s pixels. If the source format contains no alpha values, no blending will be applied. The source alpha values are not premultiplied into the source image’s other color channels.

pattern DISPLAY_PLANE_ALPHA_PER_PIXEL_PREMULTIPLIED_BIT_KHR :: DisplayPlaneAlphaFlagBitsKHR
DISPLAY_PLANE_ALPHA_PER_PIXEL_PREMULTIPLIED_BIT_KHR is equivalent to DISPLAY_PLANE_ALPHA_PER_PIXEL_BIT_KHR, except the source alpha values are assumed to be premultiplied into the source image’s other color channels.

type DisplayPlaneAlphaFlagsKHR = DisplayPlaneAlphaFlagBitsKHR

type KHR_DISPLAY_SPEC_VERSION = 23

pattern KHR_DISPLAY_SPEC_VERSION :: forall a. Integral a => a

type KHR_DISPLAY_EXTENSION_NAME = "VK_KHR_display"

pattern KHR_DISPLAY_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype DisplayKHR

VkDisplayKHR - Opaque handle to a display object

See Also

DisplayPlanePropertiesKHR, DisplayPropertiesKHR, acquireXlibDisplayEXT, createDisplayModeKHR, displayPowerControlEXT, getDisplayModeProperties2KHR, getDisplayModePropertiesKHR, getDisplayPlaneSupportedDisplaysKHR, getRandROutputDisplayEXT, registerDisplayEventEXT, releaseDisplayEXT

Constructors

DisplayKHR Word64

newtype DisplayModeKHR

VkDisplayModeKHR - Opaque handle to a display mode object

See Also

DisplayModePropertiesKHR, DisplayPlaneInfo2KHR, DisplaySurfaceCreateInfoKHR, createDisplayModeKHR, getDisplayPlaneCapabilitiesKHR

Constructors

DisplayModeKHR Word64

newtype SurfaceKHR

VkSurfaceKHR - Opaque handle to a surface object

Description

The VK_KHR_surface extension declares the SurfaceKHR object, and provides a function for destroying SurfaceKHR objects. Separate platform-specific extensions each provide a function for creating a SurfaceKHR object for the respective platform. From the application’s perspective this is an opaque handle, just like the handles of other Vulkan objects.

See Also

PhysicalDeviceSurfaceInfo2KHR, SwapchainCreateInfoKHR, createAndroidSurfaceKHR, createDirectFBSurfaceEXT, createDisplayPlaneSurfaceKHR, createHeadlessSurfaceEXT, createIOSSurfaceMVK, createImagePipeSurfaceFUCHSIA, createMacOSSurfaceMVK, createMetalSurfaceEXT, createStreamDescriptorSurfaceGGP, createViSurfaceNN, createWaylandSurfaceKHR, createWin32SurfaceKHR, createXcbSurfaceKHR, createXlibSurfaceKHR, destroySurfaceKHR, getDeviceGroupSurfacePresentModesKHR, getPhysicalDevicePresentRectanglesKHR, getPhysicalDeviceSurfaceCapabilities2EXT, getPhysicalDeviceSurfaceCapabilitiesKHR, getPhysicalDeviceSurfaceFormatsKHR, getPhysicalDeviceSurfacePresentModesKHR, getPhysicalDeviceSurfaceSupportKHR

Constructors

SurfaceKHR Word64

newtype SurfaceTransformFlagBitsKHR

VkSurfaceTransformFlagBitsKHR - presentation transforms supported on a device

See Also

CommandBufferInheritanceRenderPassTransformInfoQCOM, DisplaySurfaceCreateInfoKHR, RenderPassTransformBeginInfoQCOM, SurfaceCapabilities2EXT, SurfaceCapabilitiesKHR, SurfaceTransformFlagsKHR, SwapchainCreateInfoKHR

Constructors

SurfaceTransformFlagBitsKHR Flags

Bundled Patterns

pattern SURFACE_TRANSFORM_IDENTITY_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_IDENTITY_BIT_KHR specifies that image content is presented without being transformed.

pattern SURFACE_TRANSFORM_ROTATE_90_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_ROTATE_90_BIT_KHR specifies that image content is rotated 90 degrees clockwise.

pattern SURFACE_TRANSFORM_ROTATE_180_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_ROTATE_180_BIT_KHR specifies that image content is rotated 180 degrees clockwise.

pattern SURFACE_TRANSFORM_ROTATE_270_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_ROTATE_270_BIT_KHR specifies that image content is rotated 270 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_BIT_KHR specifies that image content is mirrored horizontally.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_90_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_90_BIT_KHR specifies that image content is mirrored horizontally, then rotated 90 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_180_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_180_BIT_KHR specifies that image content is mirrored horizontally, then rotated 180 degrees clockwise.

pattern SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_270_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_270_BIT_KHR specifies that image content is mirrored horizontally, then rotated 270 degrees clockwise.

pattern SURFACE_TRANSFORM_INHERIT_BIT_KHR :: SurfaceTransformFlagBitsKHR
SURFACE_TRANSFORM_INHERIT_BIT_KHR specifies that the presentation transform is not specified, and is instead determined by platform-specific considerations and mechanisms outside Vulkan.

type SurfaceTransformFlagsKHR = SurfaceTransformFlagBitsKHR