Vulkan.Extensions.VK_KHR_pipeline_library

Documentation

data PipelineLibraryCreateInfoKHR

VkPipelineLibraryCreateInfoKHR - Structure specifying pipeline libraries to use when creating a pipeline

Valid Usage

Each element of pLibraries must have been created with PIPELINE_CREATE_LIBRARY_BIT_KHR

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_PIPELINE_LIBRARY_CREATE_INFO_KHR

pNext must be NULL

If libraryCount is not 0, pLibraries must be a valid pointer to an array of libraryCount valid Pipeline handles

See Also

Pipeline, RayTracingPipelineCreateInfoKHR, StructureType

Constructors

PipelineLibraryCreateInfoKHR

Fields

libraries :: Vector Pipeline

pLibraries is an array of pipeline libraries to use when creating a pipeline.

type KHR_PIPELINE_LIBRARY_SPEC_VERSION = 1

pattern KHR_PIPELINE_LIBRARY_SPEC_VERSION :: forall a. Integral a => a

type KHR_PIPELINE_LIBRARY_EXTENSION_NAME = "VK_KHR_pipeline_library"

pattern KHR_PIPELINE_LIBRARY_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a