Vulkan.Extensions.VK_KHR_portability_subset

Documentation

data PhysicalDevicePortabilitySubsetFeaturesKHR

VkPhysicalDevicePortabilitySubsetFeaturesKHR - Structure describing the features that may not be supported by an implementation of the Vulkan 1.0 Portability Subset

Members

The members of the PhysicalDevicePortabilitySubsetFeaturesKHR structure describe the following features:

Description

If the PhysicalDevicePortabilitySubsetFeaturesKHR structure is included in the pNext chain of PhysicalDeviceFeatures2, it is filled with values indicating whether the features are supported. PhysicalDevicePortabilitySubsetFeaturesKHR can also be used in the pNext chain of DeviceCreateInfo to enable the features.

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

PhysicalDevicePortabilitySubsetFeaturesKHR

Fields

constantAlphaColorBlendFactors :: Bool

constantAlphaColorBlendFactors indicates whether this implementation supports constant alpha https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#framebuffer-blendfactors used as source or destination color https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#framebuffer-blending.

events :: Bool

events indicates whether this implementation supports synchronization using https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#synchronization-events.

imageViewFormatReinterpretation :: Bool

imageViewFormatReinterpretation indicates whether this implementation supports a ImageView being created with a texel format containing a different number of components, or a different number of bits in each component, than the texel format of the underlying Image.

imageViewFormatSwizzle :: Bool

imageViewFormatSwizzle indicates whether this implementation supports remapping format components using ImageViewCreateInfo::components.

imageView2DOn3DImage :: Bool

imageView2DOn3DImage indicates whether this implementation supports a Image being created with the IMAGE_CREATE_2D_ARRAY_COMPATIBLE_BIT flag set, permitting a 2D or 2D array image view to be created on a 3D Image.

multisampleArrayImage :: Bool

multisampleArrayImage indicates whether this implementation supports a Image being created as a 2D array with multiple samples per texel.

mutableComparisonSamplers :: Bool

mutableComparisonSamplers indicates whether this implementation allows descriptors with comparison samplers to be updated.

pointPolygons :: Bool

pointPolygons indicates whether this implementation supports https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#primsrast using a point https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#primsrast-polygonmode.

samplerMipLodBias :: Bool

samplerMipLodBias indicates whether this implementation supports setting a mipmap LOD bias value when creating a sampler.

separateStencilMaskRef :: Bool

separateStencilMaskRef indicates whether this implementation supports separate front and back https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#fragops-stencil reference values.

shaderSampleRateInterpolationFunctions :: Bool

shaderSampleRateInterpolationFunctions indicates whether this implementation supports fragment shaders which use the InterpolationFunction capability and the extended instructions InterpolateAtCentroid, InterpolateAtOffset, and InterpolateAtSample from the GLSL.std.450 extended instruction set. This member is only meaningful if the sampleRateShading feature is supported.

tessellationIsolines :: Bool

tessellationIsolines indicates whether this implementation supports isoline output from the https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#tessellation stage of a graphics pipeline. This member is only meaningful if tessellation shaders are supported.

tessellationPointMode :: Bool

tessellationPointMode indicates whether this implementation supports point output from the https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#tessellation stage of a graphics pipeline. This member is only meaningful if tessellation shaders are supported.

triangleFans :: Bool

triangleFans indicates whether this implementation supports https://www.khronos.org/registry/vulkan/specs/1.2-extensions/html/vkspec.html#drawing-triangle-fans primitive topology.

vertexAttributeAccessBeyondStride :: Bool

vertexAttributeAccessBeyondStride indicates whether this implementation supports accessing a vertex input attribute beyond the stride of the corresponding vertex input binding.

data PhysicalDevicePortabilitySubsetPropertiesKHR

VkPhysicalDevicePortabilitySubsetPropertiesKHR - Structure describing additional properties supported by a portable implementation

Members

The members of the PhysicalDevicePortabilitySubsetPropertiesKHR structure describe the following implementation-dependent limits:

Description

If the PhysicalDevicePortabilitySubsetPropertiesKHR structure is included in the pNext chain of PhysicalDeviceProperties2, it is filled with the implementation-dependent limits.

Valid Usage (Implicit)

See Also

StructureType

Constructors

PhysicalDevicePortabilitySubsetPropertiesKHR

Fields

minVertexInputBindingStrideAlignment :: Word32

minVertexInputBindingStrideAlignment indicates the minimum alignment for vertex input strides. VertexInputBindingDescription::stride must be a multiple of, and at least as large as, this value.

type KHR_PORTABILITY_SUBSET_SPEC_VERSION = 1

pattern KHR_PORTABILITY_SUBSET_SPEC_VERSION :: forall a. Integral a => a

type KHR_PORTABILITY_SUBSET_EXTENSION_NAME = "VK_KHR_portability_subset"

pattern KHR_PORTABILITY_SUBSET_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a