Vulkan.Extensions.VK_KHR_sampler_ycbcr_conversion

Documentation

pattern STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_CREATE_INFO_KHR :: StructureType

pattern STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_INFO_KHR :: StructureType

pattern STRUCTURE_TYPE_BIND_IMAGE_PLANE_MEMORY_INFO_KHR :: StructureType

pattern STRUCTURE_TYPE_IMAGE_PLANE_MEMORY_REQUIREMENTS_INFO_KHR :: StructureType

pattern STRUCTURE_TYPE_PHYSICAL_DEVICE_SAMPLER_YCBCR_CONVERSION_FEATURES_KHR :: StructureType

pattern STRUCTURE_TYPE_SAMPLER_YCBCR_CONVERSION_IMAGE_FORMAT_PROPERTIES_KHR :: StructureType

pattern DEBUG_REPORT_OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION_KHR_EXT :: DebugReportObjectTypeEXT

pattern OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION_KHR :: ObjectType

pattern FORMAT_G8B8G8R8_422_UNORM_KHR :: Format

pattern FORMAT_B8G8R8G8_422_UNORM_KHR :: Format

pattern FORMAT_G8_B8_R8_3PLANE_420_UNORM_KHR :: Format

pattern FORMAT_G8_B8R8_2PLANE_420_UNORM_KHR :: Format

pattern FORMAT_G8_B8_R8_3PLANE_422_UNORM_KHR :: Format

pattern FORMAT_G8_B8R8_2PLANE_422_UNORM_KHR :: Format

pattern FORMAT_G8_B8_R8_3PLANE_444_UNORM_KHR :: Format

pattern FORMAT_R10X6_UNORM_PACK16_KHR :: Format

pattern FORMAT_R10X6G10X6_UNORM_2PACK16_KHR :: Format

pattern FORMAT_R10X6G10X6B10X6A10X6_UNORM_4PACK16_KHR :: Format

pattern FORMAT_G10X6B10X6G10X6R10X6_422_UNORM_4PACK16_KHR :: Format

pattern FORMAT_B10X6G10X6R10X6G10X6_422_UNORM_4PACK16_KHR :: Format

pattern FORMAT_G10X6_B10X6_R10X6_3PLANE_420_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G10X6_B10X6R10X6_2PLANE_420_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G10X6_B10X6_R10X6_3PLANE_422_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G10X6_B10X6R10X6_2PLANE_422_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G10X6_B10X6_R10X6_3PLANE_444_UNORM_3PACK16_KHR :: Format

pattern FORMAT_R12X4_UNORM_PACK16_KHR :: Format

pattern FORMAT_R12X4G12X4_UNORM_2PACK16_KHR :: Format

pattern FORMAT_R12X4G12X4B12X4A12X4_UNORM_4PACK16_KHR :: Format

pattern FORMAT_G12X4B12X4G12X4R12X4_422_UNORM_4PACK16_KHR :: Format

pattern FORMAT_B12X4G12X4R12X4G12X4_422_UNORM_4PACK16_KHR :: Format

pattern FORMAT_G12X4_B12X4_R12X4_3PLANE_420_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G12X4_B12X4R12X4_2PLANE_420_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G12X4_B12X4_R12X4_3PLANE_422_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G12X4_B12X4R12X4_2PLANE_422_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G12X4_B12X4_R12X4_3PLANE_444_UNORM_3PACK16_KHR :: Format

pattern FORMAT_G16B16G16R16_422_UNORM_KHR :: Format

pattern FORMAT_B16G16R16G16_422_UNORM_KHR :: Format

pattern FORMAT_G16_B16_R16_3PLANE_420_UNORM_KHR :: Format

pattern FORMAT_G16_B16R16_2PLANE_420_UNORM_KHR :: Format

pattern FORMAT_G16_B16_R16_3PLANE_422_UNORM_KHR :: Format

pattern FORMAT_G16_B16R16_2PLANE_422_UNORM_KHR :: Format

pattern FORMAT_G16_B16_R16_3PLANE_444_UNORM_KHR :: Format

pattern IMAGE_ASPECT_PLANE_0_BIT_KHR :: ImageAspectFlagBits

pattern IMAGE_ASPECT_PLANE_1_BIT_KHR :: ImageAspectFlagBits

pattern IMAGE_ASPECT_PLANE_2_BIT_KHR :: ImageAspectFlagBits

pattern IMAGE_CREATE_DISJOINT_BIT_KHR :: ImageCreateFlagBits

pattern FORMAT_FEATURE_MIDPOINT_CHROMA_SAMPLES_BIT_KHR :: FormatFeatureFlagBits

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_LINEAR_FILTER_BIT_KHR :: FormatFeatureFlagBits

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_SEPARATE_RECONSTRUCTION_FILTER_BIT_KHR :: FormatFeatureFlagBits

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_BIT_KHR :: FormatFeatureFlagBits

pattern FORMAT_FEATURE_SAMPLED_IMAGE_YCBCR_CONVERSION_CHROMA_RECONSTRUCTION_EXPLICIT_FORCEABLE_BIT_KHR :: FormatFeatureFlagBits

pattern FORMAT_FEATURE_DISJOINT_BIT_KHR :: FormatFeatureFlagBits

pattern FORMAT_FEATURE_COSITED_CHROMA_SAMPLES_BIT_KHR :: FormatFeatureFlagBits

pattern SAMPLER_YCBCR_MODEL_CONVERSION_RGB_IDENTITY_KHR :: SamplerYcbcrModelConversion

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_IDENTITY_KHR :: SamplerYcbcrModelConversion

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_709_KHR :: SamplerYcbcrModelConversion

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_601_KHR :: SamplerYcbcrModelConversion

pattern SAMPLER_YCBCR_MODEL_CONVERSION_YCBCR_2020_KHR :: SamplerYcbcrModelConversion

pattern SAMPLER_YCBCR_RANGE_ITU_FULL_KHR :: SamplerYcbcrRange

pattern SAMPLER_YCBCR_RANGE_ITU_NARROW_KHR :: SamplerYcbcrRange

pattern CHROMA_LOCATION_COSITED_EVEN_KHR :: ChromaLocation

pattern CHROMA_LOCATION_MIDPOINT_KHR :: ChromaLocation

createSamplerYcbcrConversionKHR :: forall (a :: [Type]) io. (Extendss SamplerYcbcrConversionCreateInfo a, PokeChain a, MonadIO io) => Device -> SamplerYcbcrConversionCreateInfo a -> ("allocator" ::: Maybe AllocationCallbacks) -> io SamplerYcbcrConversion

destroySamplerYcbcrConversionKHR :: MonadIO io => Device -> SamplerYcbcrConversion -> ("allocator" ::: Maybe AllocationCallbacks) -> io ()

type SamplerYcbcrConversionKHR = SamplerYcbcrConversion

type SamplerYcbcrModelConversionKHR = SamplerYcbcrModelConversion

type SamplerYcbcrRangeKHR = SamplerYcbcrRange

type ChromaLocationKHR = ChromaLocation

type SamplerYcbcrConversionInfoKHR = SamplerYcbcrConversionInfo

type SamplerYcbcrConversionCreateInfoKHR = SamplerYcbcrConversionCreateInfo

type BindImagePlaneMemoryInfoKHR = BindImagePlaneMemoryInfo

type ImagePlaneMemoryRequirementsInfoKHR = ImagePlaneMemoryRequirementsInfo

type PhysicalDeviceSamplerYcbcrConversionFeaturesKHR = PhysicalDeviceSamplerYcbcrConversionFeatures

type SamplerYcbcrConversionImageFormatPropertiesKHR = SamplerYcbcrConversionImageFormatProperties

type KHR_SAMPLER_YCBCR_CONVERSION_SPEC_VERSION = 14

pattern KHR_SAMPLER_YCBCR_CONVERSION_SPEC_VERSION :: forall a. Integral a => a

type KHR_SAMPLER_YCBCR_CONVERSION_EXTENSION_NAME = "VK_KHR_sampler_ycbcr_conversion"

pattern KHR_SAMPLER_YCBCR_CONVERSION_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype DebugReportObjectTypeEXT

VkDebugReportObjectTypeEXT - Specify the type of an object handle

Description

'

DebugReportObjectTypeEXT 	 Vulkan Handle Type

DEBUG_REPORT_OBJECT_TYPE_UNKNOWN_EXT 	 Unknown/Undefined Handle

DEBUG_REPORT_OBJECT_TYPE_INSTANCE_EXT 	 Instance

DEBUG_REPORT_OBJECT_TYPE_PHYSICAL_DEVICE_EXT 	 PhysicalDevice

DEBUG_REPORT_OBJECT_TYPE_DEVICE_EXT 	 Device

DEBUG_REPORT_OBJECT_TYPE_QUEUE_EXT 	 Queue

DEBUG_REPORT_OBJECT_TYPE_SEMAPHORE_EXT 	 Semaphore

DEBUG_REPORT_OBJECT_TYPE_COMMAND_BUFFER_EXT 	 CommandBuffer

DEBUG_REPORT_OBJECT_TYPE_FENCE_EXT 	 Fence

DEBUG_REPORT_OBJECT_TYPE_DEVICE_MEMORY_EXT 	 DeviceMemory

DEBUG_REPORT_OBJECT_TYPE_BUFFER_EXT 	 Buffer

DEBUG_REPORT_OBJECT_TYPE_IMAGE_EXT 	 Image

DEBUG_REPORT_OBJECT_TYPE_EVENT_EXT 	 Event

DEBUG_REPORT_OBJECT_TYPE_QUERY_POOL_EXT 	 QueryPool

DEBUG_REPORT_OBJECT_TYPE_BUFFER_VIEW_EXT 	 BufferView

DEBUG_REPORT_OBJECT_TYPE_IMAGE_VIEW_EXT 	 ImageView

DEBUG_REPORT_OBJECT_TYPE_SHADER_MODULE_EXT 	 ShaderModule

DEBUG_REPORT_OBJECT_TYPE_PIPELINE_CACHE_EXT 	 PipelineCache

DEBUG_REPORT_OBJECT_TYPE_PIPELINE_LAYOUT_EXT 	 PipelineLayout

DEBUG_REPORT_OBJECT_TYPE_RENDER_PASS_EXT 	 RenderPass

DEBUG_REPORT_OBJECT_TYPE_PIPELINE_EXT 	 Pipeline

DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_LAYOUT_EXT 	 DescriptorSetLayout

DEBUG_REPORT_OBJECT_TYPE_SAMPLER_EXT 	 Sampler

DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_POOL_EXT 	 DescriptorPool

DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_EXT 	 DescriptorSet

DEBUG_REPORT_OBJECT_TYPE_FRAMEBUFFER_EXT 	 Framebuffer

DEBUG_REPORT_OBJECT_TYPE_COMMAND_POOL_EXT 	 CommandPool

DEBUG_REPORT_OBJECT_TYPE_SURFACE_KHR_EXT 	 SurfaceKHR

DEBUG_REPORT_OBJECT_TYPE_SWAPCHAIN_KHR_EXT 	 SwapchainKHR

DEBUG_REPORT_OBJECT_TYPE_DEBUG_REPORT_CALLBACK_EXT_EXT 	 DebugReportCallbackEXT

DEBUG_REPORT_OBJECT_TYPE_DISPLAY_KHR_EXT 	 DisplayKHR

DEBUG_REPORT_OBJECT_TYPE_DISPLAY_MODE_KHR_EXT 	 DisplayModeKHR

DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_EXT 	 DescriptorUpdateTemplate

DebugReportObjectTypeEXT and Vulkan Handle Relationship

Note

The primary expected use of ERROR_VALIDATION_FAILED_EXT is for validation layer testing. It is not expected that an application would see this error code during normal use of the validation layers.

See Also

DebugMarkerObjectNameInfoEXT, DebugMarkerObjectTagInfoEXT, debugReportMessageEXT

Constructors

DebugReportObjectTypeEXT Int32

Bundled Patterns

pattern DEBUG_REPORT_OBJECT_TYPE_UNKNOWN_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_INSTANCE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_PHYSICAL_DEVICE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DEVICE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_QUEUE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SEMAPHORE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_COMMAND_BUFFER_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_FENCE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DEVICE_MEMORY_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_BUFFER_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_IMAGE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_EVENT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_QUERY_POOL_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_BUFFER_VIEW_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_IMAGE_VIEW_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SHADER_MODULE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_PIPELINE_CACHE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_PIPELINE_LAYOUT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_RENDER_PASS_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_PIPELINE_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_LAYOUT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SAMPLER_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_POOL_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_SET_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_FRAMEBUFFER_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_COMMAND_POOL_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SURFACE_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SWAPCHAIN_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DEBUG_REPORT_CALLBACK_EXT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DISPLAY_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DISPLAY_MODE_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_VALIDATION_CACHE_EXT_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_SAMPLER_YCBCR_CONVERSION_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_ACCELERATION_STRUCTURE_KHR_EXT :: DebugReportObjectTypeEXT

pattern DEBUG_REPORT_OBJECT_TYPE_DESCRIPTOR_UPDATE_TEMPLATE_EXT :: DebugReportObjectTypeEXT