Vulkan.Extensions.VK_NV_clip_space_w_scaling

Documentation

cmdSetViewportWScalingNV

Arguments

:: forall io. MonadIO io

=> CommandBuffer
commandBuffer is the command buffer into which the command will be recorded.

-> ("firstViewport" ::: Word32)
firstViewport is the index of the first viewport whose parameters are updated by the command.

-> ("viewportWScalings" ::: Vector ViewportWScalingNV)
pViewportWScalings is a pointer to an array of ViewportWScalingNV structures specifying viewport parameters.

-> io ()

vkCmdSetViewportWScalingNV - Set the viewport W scaling on a command buffer

Description

The viewport parameters taken from element i of pViewportWScalings replace the current state for the viewport index firstViewport + i, for i in [0, viewportCount).

Valid Usage

firstViewport must be less than PhysicalDeviceLimits::maxViewports

The sum of firstViewport and viewportCount must be between 1 and PhysicalDeviceLimits::maxViewports, inclusive

Valid Usage (Implicit)

commandBuffer must be a valid CommandBuffer handle

pViewportWScalings must be a valid pointer to an array of viewportCount ViewportWScalingNV structures

commandBuffer must be in the recording state

The CommandPool that commandBuffer was allocated from must support graphics operations

viewportCount must be greater than 0

Host Synchronization

Host access to commandBuffer must be externally synchronized

Host access to the CommandPool that commandBuffer was allocated from must be externally synchronized

Command Properties

'

Command Buffer Levels 	 Render Pass Scope 	 Supported Queue Types 	 Pipeline Type

Primary Secondary 	 Both 	 Graphics

See Also

CommandBuffer, ViewportWScalingNV

data ViewportWScalingNV

VkViewportWScalingNV - Structure specifying a viewport

See Also

PipelineViewportWScalingStateCreateInfoNV, cmdSetViewportWScalingNV

Constructors

ViewportWScalingNV

Fields

xcoeff :: Float

xcoeff and ycoeff are the viewport’s W scaling factor for x and y respectively.

ycoeff :: Float

data PipelineViewportWScalingStateCreateInfoNV

VkPipelineViewportWScalingStateCreateInfoNV - Structure specifying parameters of a newly created pipeline viewport W scaling state

Valid Usage (Implicit)

See Also

Bool32, StructureType, ViewportWScalingNV

Constructors

PipelineViewportWScalingStateCreateInfoNV

Fields

viewportWScalingEnable :: Bool

viewportWScalingEnable controls whether viewport W scaling is enabled.

viewportCount :: Word32

viewportCount is the number of viewports used by W scaling, and must match the number of viewports in the pipeline if viewport W scaling is enabled.

viewportCount must be greater than 0

viewportWScalings :: Vector ViewportWScalingNV

pViewportWScalings is a pointer to an array of ViewportWScalingNV structures defining the W scaling parameters for the corresponding viewports. If the viewport W scaling state is dynamic, this member is ignored.

type NV_CLIP_SPACE_W_SCALING_SPEC_VERSION = 1

pattern NV_CLIP_SPACE_W_SCALING_SPEC_VERSION :: forall a. Integral a => a

type NV_CLIP_SPACE_W_SCALING_EXTENSION_NAME = "VK_NV_clip_space_w_scaling"

pattern NV_CLIP_SPACE_W_SCALING_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a