Vulkan.Extensions.VK_NV_dedicated_allocation

Documentation

data DedicatedAllocationImageCreateInfoNV

VkDedicatedAllocationImageCreateInfoNV - Specify that an image is bound to a dedicated memory resource

Description

Note

Using a dedicated allocation for color and depth/stencil attachments or other large images may improve performance on some devices.

Valid Usage

If dedicatedAllocation is TRUE, ImageCreateInfo::flags must not include IMAGE_CREATE_SPARSE_BINDING_BIT, IMAGE_CREATE_SPARSE_RESIDENCY_BIT, or IMAGE_CREATE_SPARSE_ALIASED_BIT

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_DEDICATED_ALLOCATION_IMAGE_CREATE_INFO_NV

See Also

Bool32, StructureType

Constructors

DedicatedAllocationImageCreateInfoNV

Fields

dedicatedAllocation :: Bool

dedicatedAllocation specifies whether the image will have a dedicated allocation bound to it.

data DedicatedAllocationBufferCreateInfoNV

VkDedicatedAllocationBufferCreateInfoNV - Specify that a buffer is bound to a dedicated memory resource

Valid Usage (Implicit)

See Also

Bool32, StructureType

Constructors

DedicatedAllocationBufferCreateInfoNV

Fields

dedicatedAllocation :: Bool

dedicatedAllocation specifies whether the buffer will have a dedicated allocation bound to it.

data DedicatedAllocationMemoryAllocateInfoNV

VkDedicatedAllocationMemoryAllocateInfoNV - Specify a dedicated memory allocation resource

Valid Usage

At least one of image and buffer must be NULL_HANDLE

If image is not NULL_HANDLE, the image must have been created with DedicatedAllocationImageCreateInfoNV::dedicatedAllocation equal to TRUE

If buffer is not NULL_HANDLE, the buffer must have been created with DedicatedAllocationBufferCreateInfoNV::dedicatedAllocation equal to TRUE

If image is not NULL_HANDLE, MemoryAllocateInfo::allocationSize must equal the MemoryRequirements::size of the image

If buffer is not NULL_HANDLE, MemoryAllocateInfo::allocationSize must equal the MemoryRequirements::size of the buffer

If image is not NULL_HANDLE and MemoryAllocateInfo defines a memory import operation, the memory being imported must also be a dedicated image allocation and image must be identical to the image associated with the imported memory

If buffer is not NULL_HANDLE and MemoryAllocateInfo defines a memory import operation, the memory being imported must also be a dedicated buffer allocation and buffer must be identical to the buffer associated with the imported memory

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_DEDICATED_ALLOCATION_MEMORY_ALLOCATE_INFO_NV

If image is not NULL_HANDLE, image must be a valid Image handle

If buffer is not NULL_HANDLE, buffer must be a valid Buffer handle

Both of buffer, and image that are valid handles of non-ignored parameters must have been created, allocated, or retrieved from the same Device

See Also

Buffer, Image, StructureType

Constructors

DedicatedAllocationMemoryAllocateInfoNV

Fields

image :: Image

image is NULL_HANDLE or a handle of an image which this memory will be bound to.

buffer :: Buffer

buffer is NULL_HANDLE or a handle of a buffer which this memory will be bound to.

type NV_DEDICATED_ALLOCATION_SPEC_VERSION = 1

pattern NV_DEDICATED_ALLOCATION_SPEC_VERSION :: forall a. Integral a => a

type NV_DEDICATED_ALLOCATION_EXTENSION_NAME = "VK_NV_dedicated_allocation"

pattern NV_DEDICATED_ALLOCATION_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a