Vulkan.Extensions.VK_NV_external_memory

Documentation

data ExternalMemoryImageCreateInfoNV

VkExternalMemoryImageCreateInfoNV - Specify that an image may be backed by external memory

Valid Usage (Implicit)

See Also

ExternalMemoryHandleTypeFlagsNV, StructureType

Constructors

ExternalMemoryImageCreateInfoNV

Fields

handleTypes :: ExternalMemoryHandleTypeFlagsNV

handleTypes is a bitmask of ExternalMemoryHandleTypeFlagBitsNV specifying one or more external memory handle types.

handleTypes must be a valid combination of ExternalMemoryHandleTypeFlagBitsNV values

data ExportMemoryAllocateInfoNV

VkExportMemoryAllocateInfoNV - Specify memory handle types that may be exported

Valid Usage (Implicit)

See Also

ExternalMemoryHandleTypeFlagsNV, StructureType

Constructors

ExportMemoryAllocateInfoNV

Fields

handleTypes :: ExternalMemoryHandleTypeFlagsNV

handleTypes is a bitmask of ExternalMemoryHandleTypeFlagBitsNV specifying one or more memory handle types that may be exported. Multiple handle types may be requested for the same allocation as long as they are compatible, as reported by getPhysicalDeviceExternalImageFormatPropertiesNV.

handleTypes must be a valid combination of ExternalMemoryHandleTypeFlagBitsNV values

type NV_EXTERNAL_MEMORY_SPEC_VERSION = 1

pattern NV_EXTERNAL_MEMORY_SPEC_VERSION :: forall a. Integral a => a

type NV_EXTERNAL_MEMORY_EXTENSION_NAME = "VK_NV_external_memory"

pattern NV_EXTERNAL_MEMORY_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

newtype ExternalMemoryHandleTypeFlagBitsNV

VkExternalMemoryHandleTypeFlagBitsNV - Bitmask specifying external memory handle types

See Also

ExternalMemoryHandleTypeFlagsNV

Constructors

ExternalMemoryHandleTypeFlagBitsNV Flags

Bundled Patterns

pattern EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT_NV specifies a handle to memory returned by getMemoryWin32HandleNV, or one duplicated from such a handle using DuplicateHandle().

pattern EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_NV specifies a handle to memory returned by getMemoryWin32HandleNV.

pattern EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_BIT_NV specifies a valid NT handle to memory returned by IDXGIResource1::CreateSharedHandle, or a handle duplicated from such a handle using DuplicateHandle().

pattern EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_KMT_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_KMT_BIT_NV specifies a handle to memory returned by IDXGIResource::GetSharedHandle().

type ExternalMemoryHandleTypeFlagsNV = ExternalMemoryHandleTypeFlagBitsNV