Vulkan.Extensions.VK_NV_external_memory_capabilities

Documentation

getPhysicalDeviceExternalImageFormatPropertiesNV

Arguments

:: forall io. MonadIO io

=> PhysicalDevice
physicalDevice is the physical device from which to query the image capabilities

physicalDevice must be a valid PhysicalDevice handle

-> Format
format is the image format, corresponding to ImageCreateInfo::format.

format must be a valid Format value

-> ImageType
type is the image type, corresponding to ImageCreateInfo::imageType.

type must be a valid ImageType value

-> ImageTiling
tiling is the image tiling, corresponding to ImageCreateInfo::tiling.

tiling must be a valid ImageTiling value

-> ImageUsageFlags
usage is the intended usage of the image, corresponding to ImageCreateInfo::usage.

usage must be a valid combination of ImageUsageFlagBits values

usage must not be 0

-> ImageCreateFlags
flags is a bitmask describing additional parameters of the image, corresponding to ImageCreateInfo::flags.

flags must be a valid combination of ImageCreateFlagBits values

-> ("externalHandleType" ::: ExternalMemoryHandleTypeFlagsNV)
externalHandleType is either one of the bits from ExternalMemoryHandleTypeFlagBitsNV, or 0.

externalHandleType must be a valid combination of ExternalMemoryHandleTypeFlagBitsNV values

-> io ExternalImageFormatPropertiesNV

vkGetPhysicalDeviceExternalImageFormatPropertiesNV - determine image capabilities compatible with external memory handle types

Description

If externalHandleType is 0, pExternalImageFormatProperties->imageFormatProperties will return the same values as a call to getPhysicalDeviceImageFormatProperties, and the other members of pExternalImageFormatProperties will all be 0. Otherwise, they are filled in as described for ExternalImageFormatPropertiesNV.

Return Codes

Success

SUCCESS

Failure

ERROR_OUT_OF_HOST_MEMORY

ERROR_OUT_OF_DEVICE_MEMORY

ERROR_FORMAT_NOT_SUPPORTED

See Also

ExternalImageFormatPropertiesNV, ExternalMemoryHandleTypeFlagsNV, Format, ImageCreateFlags, ImageTiling, ImageType, ImageUsageFlags, PhysicalDevice

data ExternalImageFormatPropertiesNV

VkExternalImageFormatPropertiesNV - Structure specifying external image format properties

See Also

ExternalMemoryFeatureFlagsNV, ExternalMemoryHandleTypeFlagsNV, ImageFormatProperties, getPhysicalDeviceExternalImageFormatPropertiesNV

Constructors

ExternalImageFormatPropertiesNV

Fields

imageFormatProperties :: ImageFormatProperties

imageFormatProperties will be filled in as when calling getPhysicalDeviceImageFormatProperties, but the values returned may vary depending on the external handle type requested.

externalMemoryFeatures :: ExternalMemoryFeatureFlagsNV

externalMemoryFeatures is a bitmask of ExternalMemoryFeatureFlagBitsNV, indicating properties of the external memory handle type (getPhysicalDeviceExternalImageFormatPropertiesNV::externalHandleType) being queried, or 0 if the external memory handle type is 0.

exportFromImportedHandleTypes :: ExternalMemoryHandleTypeFlagsNV

exportFromImportedHandleTypes is a bitmask of ExternalMemoryHandleTypeFlagBitsNV containing a bit set for every external handle type that may be used to create memory from which the handles of the type specified in getPhysicalDeviceExternalImageFormatPropertiesNV::externalHandleType can be exported, or 0 if the external memory handle type is 0.

compatibleHandleTypes :: ExternalMemoryHandleTypeFlagsNV

compatibleHandleTypes is a bitmask of ExternalMemoryHandleTypeFlagBitsNV containing a bit set for every external handle type that may be specified simultaneously with the handle type specified by getPhysicalDeviceExternalImageFormatPropertiesNV::externalHandleType when calling allocateMemory, or 0 if the external memory handle type is 0. compatibleHandleTypes will always contain getPhysicalDeviceExternalImageFormatPropertiesNV::externalHandleType

newtype ExternalMemoryHandleTypeFlagBitsNV

VkExternalMemoryHandleTypeFlagBitsNV - Bitmask specifying external memory handle types

See Also

ExternalMemoryHandleTypeFlagsNV

Constructors

ExternalMemoryHandleTypeFlagBitsNV Flags

Bundled Patterns

pattern EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT_NV specifies a handle to memory returned by getMemoryWin32HandleNV, or one duplicated from such a handle using DuplicateHandle().

pattern EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_NV specifies a handle to memory returned by getMemoryWin32HandleNV.

pattern EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_BIT_NV specifies a valid NT handle to memory returned by IDXGIResource1::CreateSharedHandle, or a handle duplicated from such a handle using DuplicateHandle().

pattern EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_KMT_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_KMT_BIT_NV specifies a handle to memory returned by IDXGIResource::GetSharedHandle().

type ExternalMemoryHandleTypeFlagsNV = ExternalMemoryHandleTypeFlagBitsNV

newtype ExternalMemoryFeatureFlagBitsNV

VkExternalMemoryFeatureFlagBitsNV - Bitmask specifying external memory features

See Also

ExternalImageFormatPropertiesNV, ExternalMemoryFeatureFlagsNV, getPhysicalDeviceExternalImageFormatPropertiesNV

Constructors

ExternalMemoryFeatureFlagBitsNV Flags

Bundled Patterns

pattern EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT_NV :: ExternalMemoryFeatureFlagBitsNV
EXTERNAL_MEMORY_FEATURE_DEDICATED_ONLY_BIT_NV specifies that external memory of the specified type must be created as a dedicated allocation when used in the manner specified.

pattern EXTERNAL_MEMORY_FEATURE_EXPORTABLE_BIT_NV :: ExternalMemoryFeatureFlagBitsNV
EXTERNAL_MEMORY_FEATURE_EXPORTABLE_BIT_NV specifies that the implementation supports exporting handles of the specified type.

pattern EXTERNAL_MEMORY_FEATURE_IMPORTABLE_BIT_NV :: ExternalMemoryFeatureFlagBitsNV
EXTERNAL_MEMORY_FEATURE_IMPORTABLE_BIT_NV specifies that the implementation supports importing handles of the specified type.

type ExternalMemoryFeatureFlagsNV = ExternalMemoryFeatureFlagBitsNV

type NV_EXTERNAL_MEMORY_CAPABILITIES_SPEC_VERSION = 1

pattern NV_EXTERNAL_MEMORY_CAPABILITIES_SPEC_VERSION :: forall a. Integral a => a

type NV_EXTERNAL_MEMORY_CAPABILITIES_EXTENSION_NAME = "VK_NV_external_memory_capabilities"

pattern NV_EXTERNAL_MEMORY_CAPABILITIES_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a