Vulkan.Extensions.VK_NV_external_memory_win32

Documentation

getMemoryWin32HandleNV

Arguments

:: forall io. MonadIO io

=> Device
device is the logical device that owns the memory.

device must be a valid Device handle

-> DeviceMemory
memory is the DeviceMemory object.

memory must be a valid DeviceMemory handle

memory must have been created, allocated, or retrieved from device

-> ExternalMemoryHandleTypeFlagsNV
handleType is a bitmask of ExternalMemoryHandleTypeFlagBitsNV containing a single bit specifying the type of handle requested.

handleType must be a flag specified in ExportMemoryAllocateInfoNV::handleTypes when allocating memory

handleType must be a valid combination of ExternalMemoryHandleTypeFlagBitsNV values

handleType must not be 0

-> io HANDLE

vkGetMemoryWin32HandleNV - retrieve Win32 handle to a device memory object

Return Codes

Success

SUCCESS

Failure

ERROR_TOO_MANY_OBJECTS

ERROR_OUT_OF_HOST_MEMORY

See Also

Device, DeviceMemory, ExternalMemoryHandleTypeFlagsNV

data ImportMemoryWin32HandleInfoNV

VkImportMemoryWin32HandleInfoNV - import Win32 memory created on the same physical device

Description

If handleType is 0, this structure is ignored by consumers of the MemoryAllocateInfo structure it is chained from.

Valid Usage (Implicit)

See Also

ExternalMemoryHandleTypeFlagsNV, StructureType

Constructors

ImportMemoryWin32HandleInfoNV

Fields

handleType :: ExternalMemoryHandleTypeFlagsNV

handleType is 0 or a ExternalMemoryHandleTypeFlagBitsNV value specifying the type of memory handle in handle.

handleType must not have more than one bit set

handleType must be a valid combination of ExternalMemoryHandleTypeFlagBitsNV values

handle :: HANDLE

handle is a Windows HANDLE referring to the memory.

handle must be a valid handle to memory, obtained as specified by handleType

data ExportMemoryWin32HandleInfoNV

VkExportMemoryWin32HandleInfoNV - specify security attributes and access rights for Win32 memory handles

Description

If this structure is not present, or if pAttributes is set to NULL, default security descriptor values will be used, and child processes created by the application will not inherit the handle, as described in the MSDN documentation for “Synchronization Object Security and Access Rights”1. Further, if the structure is not present, the access rights will be

DXGI_SHARED_RESOURCE_READ | DXGI_SHARED_RESOURCE_WRITE

1

https://docs.microsoft.com/en-us/windows/win32/sync/synchronization-object-security-and-access-rights

Valid Usage (Implicit)

sType must be STRUCTURE_TYPE_EXPORT_MEMORY_WIN32_HANDLE_INFO_NV

If pAttributes is not NULL, pAttributes must be a valid pointer to a valid SECURITY_ATTRIBUTES value

See Also

StructureType

Constructors

ExportMemoryWin32HandleInfoNV

Fields

attributes :: Ptr SECURITY_ATTRIBUTES

pAttributes is a pointer to a Windows SECURITY_ATTRIBUTES structure specifying security attributes of the handle.

dwAccess :: DWORD

dwAccess is a DWORD specifying access rights of the handle.

type NV_EXTERNAL_MEMORY_WIN32_SPEC_VERSION = 1

pattern NV_EXTERNAL_MEMORY_WIN32_SPEC_VERSION :: forall a. Integral a => a

type NV_EXTERNAL_MEMORY_WIN32_EXTENSION_NAME = "VK_NV_external_memory_win32"

pattern NV_EXTERNAL_MEMORY_WIN32_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a

type HANDLE = Ptr ()

type DWORD = Word32

data SECURITY_ATTRIBUTES

newtype ExternalMemoryHandleTypeFlagBitsNV

VkExternalMemoryHandleTypeFlagBitsNV - Bitmask specifying external memory handle types

See Also

ExternalMemoryHandleTypeFlagsNV

Constructors

ExternalMemoryHandleTypeFlagBitsNV Flags

Bundled Patterns

pattern EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_BIT_NV specifies a handle to memory returned by getMemoryWin32HandleNV, or one duplicated from such a handle using DuplicateHandle().

pattern EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_OPAQUE_WIN32_KMT_BIT_NV specifies a handle to memory returned by getMemoryWin32HandleNV.

pattern EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_BIT_NV specifies a valid NT handle to memory returned by IDXGIResource1::CreateSharedHandle, or a handle duplicated from such a handle using DuplicateHandle().

pattern EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_KMT_BIT_NV :: ExternalMemoryHandleTypeFlagBitsNV
EXTERNAL_MEMORY_HANDLE_TYPE_D3D11_IMAGE_KMT_BIT_NV specifies a handle to memory returned by IDXGIResource::GetSharedHandle().

type ExternalMemoryHandleTypeFlagsNV = ExternalMemoryHandleTypeFlagBitsNV