Vulkan.Extensions.VK_NV_viewport_swizzle

Documentation

data ViewportSwizzleNV

VkViewportSwizzleNV - Structure specifying a viewport swizzle

Valid Usage (Implicit)

See Also

PipelineViewportSwizzleStateCreateInfoNV, ViewportCoordinateSwizzleNV

Constructors

ViewportSwizzleNV

Fields

x :: ViewportCoordinateSwizzleNV

x must be a valid ViewportCoordinateSwizzleNV value

y :: ViewportCoordinateSwizzleNV

y must be a valid ViewportCoordinateSwizzleNV value

z :: ViewportCoordinateSwizzleNV

z must be a valid ViewportCoordinateSwizzleNV value

w :: ViewportCoordinateSwizzleNV

w must be a valid ViewportCoordinateSwizzleNV value

data PipelineViewportSwizzleStateCreateInfoNV

VkPipelineViewportSwizzleStateCreateInfoNV - Structure specifying swizzle applied to primitive clip coordinates

Valid Usage (Implicit)

See Also

PipelineViewportSwizzleStateCreateFlagsNV, StructureType, ViewportSwizzleNV

Constructors

PipelineViewportSwizzleStateCreateInfoNV

Fields

flags :: PipelineViewportSwizzleStateCreateFlagsNV

flags must be 0

viewportSwizzles :: Vector ViewportSwizzleNV

pViewportSwizzles must be a valid pointer to an array of viewportCount valid ViewportSwizzleNV structures

newtype PipelineViewportSwizzleStateCreateFlagsNV

VkPipelineViewportSwizzleStateCreateFlagsNV - Reserved for future use

Description

PipelineViewportSwizzleStateCreateFlagsNV is a bitmask type for setting a mask, but is currently reserved for future use.

See Also

PipelineViewportSwizzleStateCreateInfoNV

Constructors

PipelineViewportSwizzleStateCreateFlagsNV Flags

newtype ViewportCoordinateSwizzleNV

VkViewportCoordinateSwizzleNV - Specify how a viewport coordinate is swizzled

Description

These values are described in detail in Viewport Swizzle.

See Also

ViewportSwizzleNV

Constructors

ViewportCoordinateSwizzleNV Int32

Bundled Patterns

pattern VIEWPORT_COORDINATE_SWIZZLE_POSITIVE_X_NV :: ViewportCoordinateSwizzleNV

pattern VIEWPORT_COORDINATE_SWIZZLE_NEGATIVE_X_NV :: ViewportCoordinateSwizzleNV

pattern VIEWPORT_COORDINATE_SWIZZLE_POSITIVE_Y_NV :: ViewportCoordinateSwizzleNV

pattern VIEWPORT_COORDINATE_SWIZZLE_NEGATIVE_Y_NV :: ViewportCoordinateSwizzleNV

pattern VIEWPORT_COORDINATE_SWIZZLE_POSITIVE_Z_NV :: ViewportCoordinateSwizzleNV

pattern VIEWPORT_COORDINATE_SWIZZLE_NEGATIVE_Z_NV :: ViewportCoordinateSwizzleNV

pattern VIEWPORT_COORDINATE_SWIZZLE_POSITIVE_W_NV :: ViewportCoordinateSwizzleNV

pattern VIEWPORT_COORDINATE_SWIZZLE_NEGATIVE_W_NV :: ViewportCoordinateSwizzleNV

type NV_VIEWPORT_SWIZZLE_SPEC_VERSION = 1

pattern NV_VIEWPORT_SWIZZLE_SPEC_VERSION :: forall a. Integral a => a

type NV_VIEWPORT_SWIZZLE_EXTENSION_NAME = "VK_NV_viewport_swizzle"

pattern NV_VIEWPORT_SWIZZLE_EXTENSION_NAME :: forall a. (Eq a, IsString a) => a